Aiops

This page documents function available when using the Aiops module, created with @service Aiops.

Index

Main.Aiops.create_investigation_group

Main.Aiops.delete_investigation_group

Main.Aiops.delete_investigation_group_policy

Main.Aiops.get_investigation_group

Main.Aiops.get_investigation_group_policy

Main.Aiops.list_investigation_groups

Main.Aiops.list_tags_for_resource

Main.Aiops.put_investigation_group_policy

Main.Aiops.tag_resource

Main.Aiops.untag_resource

Main.Aiops.update_investigation_group

Documentation

Main.Aiops.create_investigation_group — Function

create_investigation_group(name, role_arn)
create_investigation_group(name, role_arn, params::Dict{String,<:Any})

Creates an investigation group in your account. Creating an investigation group is a one- time setup task for each Region in your account. It is a necessary task to be able to perform investigations.

Settings in the investigation group help you centrally manage the common properties of your investigations, such as the following:

Who can access the investigations

Whether investigation data is encrypted with a customer managed Key Management Service key.

How long investigations and their data are retained by default.

Currently, you can have one investigation group in each Region in your account. Each investigation in a Region is a part of the investigation group in that Region

To create an investigation group and set up CloudWatch investigations, you must be signed in to an IAM principal that has either the AIOpsConsoleAdminPolicy or the AdministratorAccess IAM policy attached, or to an account that has similar permissions.

Important

You can configure CloudWatch alarms to start investigations and add events to investigations. If you create your investigation group with CreateInvestigationGroup and you want to enable alarms to do this, you must use PutInvestigationGroupPolicy to create a resource policy that grants this permission to CloudWatch alarms.

For more information about configuring CloudWatch alarms, see Using Amazon CloudWatch alarms

Arguments

name: Provides a name for the investigation group.

role_arn: Specify the ARN of the IAM role that CloudWatch investigations will use when it gathers investigation data. The permissions in this role determine which of your resources that CloudWatch investigations will have access to during investigations.

For more information, see How to control what data CloudWatch investigations has access to during investigations.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"chatbotNotificationChannel": Use this structure to integrate CloudWatch investigations with chat applications. This structure is a string array. For the first string, specify the ARN of an Amazon SNS topic. For the array of strings, specify the ARNs of one or more chat applications configurations that you want to associate with that topic. For more information about these configuration ARNs, see Getting started with Amazon Q in chat applications and Resource type defined by Amazon Web Services Chatbot.

"crossAccountConfigurations": List of sourceRoleArn values that have been configured for cross-account access.

"encryptionConfiguration": Use this structure if you want to use a customer managed KMS key to encrypt your investigation data. If you omit this parameter, CloudWatch investigations will use an Amazon Web Services key to encrypt the data. For more information, see Encryption of investigation data.

"isCloudTrailEventHistoryEnabled": Specify true to enable CloudWatch investigations to have access to change events that are recorded by CloudTrail. The default is true.

"retentionInDays": Specify how long that investigation data is kept. For more information, see Operational investigation data retention.

If you omit this parameter, the default of 90 days is used.

"tagKeyBoundaries": Enter the existing custom tag keys for custom applications in your system. Resource tags help CloudWatch investigations narrow the search space when it is unable to discover definite relationships between resources. For example, to discover that an Amazon ECS service depends on an Amazon RDS database, CloudWatch investigations can discover this relationship using data sources such as X-Ray and CloudWatch Application Signals. However, if you haven't deployed these features, CloudWatch investigations will attempt to identify possible relationships. Tag boundaries can be used to narrow the resources that will be discovered by CloudWatch investigations in these cases.

You don't need to enter tags created by myApplications or CloudFormation, because CloudWatch investigations can automatically detect those tags.

"tags": A list of key-value pairs to associate with the investigation group. You can associate as many as 50 tags with an investigation group. To be able to associate tags when you create the investigation group, you must have the cloudwatch:TagResource permission.

Tags can help you organize and categorize your resources. You can also use them to scope user permissions by granting a user permission to access or change only resources with certain tag values.

Main.Aiops.delete_investigation_group — Function

delete_investigation_group(identifier)
delete_investigation_group(identifier, params::Dict{String,<:Any})

Deletes the specified investigation group from your account. You can currently have one investigation group per Region in your account. After you delete an investigation group, you can later create a new investigation group in the same Region.

Arguments

identifier: Specify either the name or the ARN of the investigation group that you want to delete.

Main.Aiops.delete_investigation_group_policy — Function

delete_investigation_group_policy(identifier)
delete_investigation_group_policy(identifier, params::Dict{String,<:Any})

Removes the IAM resource policy from being associated with the investigation group that you specify.

Arguments

identifier: Specify either the name or the ARN of the investigation group that you want to remove the policy from.

Main.Aiops.get_investigation_group — Function

get_investigation_group(identifier)
get_investigation_group(identifier, params::Dict{String,<:Any})

Returns the configuration information for the specified investigation group.

Arguments

identifier: Specify either the name or the ARN of the investigation group that you want to view. This is used to set the name of the investigation group.

Main.Aiops.get_investigation_group_policy — Function

get_investigation_group_policy(identifier)
get_investigation_group_policy(identifier, params::Dict{String,<:Any})

Returns the JSON of the IAM resource policy associated with the specified investigation group in a string. For example, {\"Version\":\"2012-10-17\",\"Statement\":[{\"Effect\":\"Allow\",\"Principal\":{\"Service\":\"aiops.alarms.cloudwatch.amazonaws.com\"},\"Action\":[\"aiops:CreateInvestigation\",\"aiops:CreateInvestigationEvent\"],\"Resource\":\"*\",\"Condition\":{\"StringEquals\":{\"aws:SourceAccount\":\"111122223333\"},\"ArnLike\":{\"aws:SourceArn\":\"arn:aws:cloudwatch:us-east-1:111122223333:alarm:*\"}}}]}.

Arguments

identifier: Specify either the name or the ARN of the investigation group that you want to view the policy of.

Main.Aiops.list_investigation_groups — Function

list_investigation_groups()
list_investigation_groups(params::Dict{String,<:Any})

Returns the ARN and name of each investigation group in the account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return in one operation. If you omit this parameter, the default of 50 is used.

"nextToken": Include this value, if it was returned by the previous operation, to get the next set of service operations.

Main.Aiops.list_tags_for_resource — Function

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Displays the tags associated with a CloudWatch investigations resource. Currently, investigation groups support tagging.

Arguments

resource_arn: The ARN of the CloudWatch investigations resource that you want to view tags for. You can use the ListInvestigationGroups operation to find the ARNs of investigation groups.

The ARN format for an investigation group is arn:aws:aiops:*Region*:*account-id*:investigation-group:*investigation-group-id*.

Main.Aiops.put_investigation_group_policy — Function

put_investigation_group_policy(identifier, policy)
put_investigation_group_policy(identifier, policy, params::Dict{String,<:Any})

Creates an IAM resource policy and assigns it to the specified investigation group.

If you create your investigation group with CreateInvestigationGroup and you want to enable CloudWatch alarms to create investigations and add events to investigations, you must use this operation to create a policy similar to this example.

{ "Version": "2008-10-17", "Statement": [ { "Effect": "Allow", "Principal": { "Service": "aiops.alarms.cloudwatch.amazonaws.com" }, "Action": [ "aiops:CreateInvestigation", "aiops:CreateInvestigationEvent" ], "Resource": "*", "Condition": { "StringEquals": { "aws:SourceAccount": "account-id" }, "ArnLike": { "aws:SourceArn": "arn:aws:cloudwatch:region:account-id:alarm:*" } } } ] }

Arguments

identifier: Specify either the name or the ARN of the investigation group that you want to assign the policy to.

policy: The policy, in JSON format.

Main.Aiops.tag_resource — Function

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Assigns one or more tags (key-value pairs) to the specified resource.

Tags can help you organize and categorize your resources. You can also use them to scope user permissions by granting a user permission to access or change only resources with certain tag values.

Tags don't have any semantic meaning to Amazon Web Services and are interpreted strictly as strings of characters.

You can associate as many as 50 tags with a resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource that you want to apply the tags to. You can use the ListInvestigationGroups operation to find the ARNs of investigation groups.

tags: The list of key-value pairs to associate with the resource.

Main.Aiops.untag_resource — Function

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes one or more tags from the specified resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource that you want to remove the tags from. You can use theListInvestigationGroups operation to find the ARNs of investigation groups.

tag_keys: The list of tag keys to remove from the resource.

Main.Aiops.update_investigation_group — Function

update_investigation_group(identifier)
update_investigation_group(identifier, params::Dict{String,<:Any})

Updates the configuration of the specified investigation group.

Arguments

identifier: Specify either the name or the ARN of the investigation group that you want to modify.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"chatbotNotificationChannel": Use this structure to integrate CloudWatch investigations with chat applications. This structure is a string array. For the first string, specify the ARN of an Amazon SNS topic. For the array of strings, specify the ARNs of one or more chat applications configurations that you want to associate with that topic. For more information about these configuration ARNs, see Getting started with Amazon Q in chat applications and Resource type defined by Amazon Web Services Chatbot.

"crossAccountConfigurations": Used to configure cross-account access for an investigation group. It allows the investigation group to access resources in other accounts.

"encryptionConfiguration": Use this structure if you want to use a customer managed KMS key to encrypt your investigation data. If you omit this parameter, CloudWatch investigations will use an Amazon Web Services key to encrypt the data. For more information, see Encryption of investigation data.

"isCloudTrailEventHistoryEnabled": Specify true to enable CloudWatch investigations to have access to change events that are recorded by CloudTrail. The default is true.

"roleArn": Specify this field if you want to change the IAM role that CloudWatch investigations will use when it gathers investigation data. To do so, specify the ARN of the new role.

The permissions in this role determine which of your resources that CloudWatch investigations will have access to during investigations.

For more information, see How to control what data CloudWatch investigations has access to during investigations.

"tagKeyBoundaries": Enter the existing custom tag keys for custom applications in your system. Resource tags help CloudWatch investigations narrow the search space when it is unable to discover definite relationships between resources. For example, to discover that an Amazon ECS service depends on an Amazon RDS database, CloudWatch investigations can discover this relationship using data sources such as X-Ray and CloudWatch Application Signals. However, if you haven't deployed these features, CloudWatch investigations will attempt to identify possible relationships. Tag boundaries can be used to narrow the resources that will be discovered by CloudWatch investigations in these cases.

You don't need to enter tags created by myApplications or CloudFormation, because CloudWatch investigations can automatically detect those tags.