Alexa For Business

This page documents function available when using the Alexa_For_Business module, created with @service Alexa_For_Business.

Index

Main.Alexa_For_Business.approve_skill

Main.Alexa_For_Business.associate_contact_with_address_book

Main.Alexa_For_Business.associate_device_with_network_profile

Main.Alexa_For_Business.associate_device_with_room

Main.Alexa_For_Business.associate_skill_group_with_room

Main.Alexa_For_Business.associate_skill_with_skill_group

Main.Alexa_For_Business.associate_skill_with_users

Main.Alexa_For_Business.create_address_book

Main.Alexa_For_Business.create_business_report_schedule

Main.Alexa_For_Business.create_conference_provider

Main.Alexa_For_Business.create_contact

Main.Alexa_For_Business.create_gateway_group

Main.Alexa_For_Business.create_network_profile

Main.Alexa_For_Business.create_profile

Main.Alexa_For_Business.create_room

Main.Alexa_For_Business.create_skill_group

Main.Alexa_For_Business.create_user

Main.Alexa_For_Business.delete_address_book

Main.Alexa_For_Business.delete_business_report_schedule

Main.Alexa_For_Business.delete_conference_provider

Main.Alexa_For_Business.delete_contact

Main.Alexa_For_Business.delete_device

Main.Alexa_For_Business.delete_device_usage_data

Main.Alexa_For_Business.delete_gateway_group

Main.Alexa_For_Business.delete_network_profile

Main.Alexa_For_Business.delete_profile

Main.Alexa_For_Business.delete_room

Main.Alexa_For_Business.delete_room_skill_parameter

Main.Alexa_For_Business.delete_skill_authorization

Main.Alexa_For_Business.delete_skill_group

Main.Alexa_For_Business.delete_user

Main.Alexa_For_Business.disassociate_contact_from_address_book

Main.Alexa_For_Business.disassociate_device_from_room

Main.Alexa_For_Business.disassociate_skill_from_skill_group

Main.Alexa_For_Business.disassociate_skill_from_users

Main.Alexa_For_Business.disassociate_skill_group_from_room

Main.Alexa_For_Business.forget_smart_home_appliances

Main.Alexa_For_Business.get_address_book

Main.Alexa_For_Business.get_conference_preference

Main.Alexa_For_Business.get_conference_provider

Main.Alexa_For_Business.get_contact

Main.Alexa_For_Business.get_device

Main.Alexa_For_Business.get_gateway

Main.Alexa_For_Business.get_gateway_group

Main.Alexa_For_Business.get_invitation_configuration

Main.Alexa_For_Business.get_network_profile

Main.Alexa_For_Business.get_profile

Main.Alexa_For_Business.get_room

Main.Alexa_For_Business.get_room_skill_parameter

Main.Alexa_For_Business.get_skill_group

Main.Alexa_For_Business.list_business_report_schedules

Main.Alexa_For_Business.list_conference_providers

Main.Alexa_For_Business.list_device_events

Main.Alexa_For_Business.list_gateway_groups

Main.Alexa_For_Business.list_gateways

Main.Alexa_For_Business.list_skills

Main.Alexa_For_Business.list_skills_store_categories

Main.Alexa_For_Business.list_skills_store_skills_by_category

Main.Alexa_For_Business.list_smart_home_appliances

Main.Alexa_For_Business.list_tags

Main.Alexa_For_Business.put_conference_preference

Main.Alexa_For_Business.put_invitation_configuration

Main.Alexa_For_Business.put_room_skill_parameter

Main.Alexa_For_Business.put_skill_authorization

Main.Alexa_For_Business.register_avsdevice

Main.Alexa_For_Business.reject_skill

Main.Alexa_For_Business.resolve_room

Main.Alexa_For_Business.revoke_invitation

Main.Alexa_For_Business.search_address_books

Main.Alexa_For_Business.search_contacts

Main.Alexa_For_Business.search_devices

Main.Alexa_For_Business.search_network_profiles

Main.Alexa_For_Business.search_profiles

Main.Alexa_For_Business.search_rooms

Main.Alexa_For_Business.search_skill_groups

Main.Alexa_For_Business.search_users

Main.Alexa_For_Business.send_announcement

Main.Alexa_For_Business.send_invitation

Main.Alexa_For_Business.start_device_sync

Main.Alexa_For_Business.start_smart_home_appliance_discovery

Main.Alexa_For_Business.tag_resource

Main.Alexa_For_Business.untag_resource

Main.Alexa_For_Business.update_address_book

Main.Alexa_For_Business.update_business_report_schedule

Main.Alexa_For_Business.update_conference_provider

Main.Alexa_For_Business.update_contact

Main.Alexa_For_Business.update_device

Main.Alexa_For_Business.update_gateway

Main.Alexa_For_Business.update_gateway_group

Main.Alexa_For_Business.update_network_profile

Main.Alexa_For_Business.update_profile

Main.Alexa_For_Business.update_room

Main.Alexa_For_Business.update_skill_group

Documentation

Main.Alexa_For_Business.approve_skill — Method

ApproveSkill()

Associates a skill with the organization under the customer's AWS account. If a skill is private, the user implicitly accepts access to this skill during enablement.

Required Parameters

SkillId: The unique identifier of the skill.

Main.Alexa_For_Business.associate_contact_with_address_book — Method

AssociateContactWithAddressBook()

Associates a contact with a given address book.

Required Parameters

AddressBookArn: The ARN of the address book with which to associate the contact.

ContactArn: The ARN of the contact to associate with an address book.

Main.Alexa_For_Business.associate_device_with_network_profile — Method

AssociateDeviceWithNetworkProfile()

Associates a device with the specified network profile.

Required Parameters

DeviceArn: The device ARN.

NetworkProfileArn: The ARN of the network profile to associate with a device.

Main.Alexa_For_Business.associate_device_with_room — Method

AssociateDeviceWithRoom()

Associates a device with a given room. This applies all the settings from the room profile to the device, and all the skills in any skill groups added to that room. This operation requires the device to be online, or else a manual sync is required.

Optional Parameters

DeviceArn: The ARN of the device to associate to a room. Required.

RoomArn: The ARN of the room with which to associate the device. Required.

Main.Alexa_For_Business.associate_skill_group_with_room — Method

AssociateSkillGroupWithRoom()

Associates a skill group with a given room. This enables all skills in the associated skill group on all devices in the room.

Optional Parameters

RoomArn: The ARN of the room with which to associate the skill group. Required.

SkillGroupArn: The ARN of the skill group to associate with a room. Required.

Main.Alexa_For_Business.associate_skill_with_skill_group — Method

AssociateSkillWithSkillGroup()

Associates a skill with a skill group.

Required Parameters

SkillId: The unique identifier of the skill.

Optional Parameters

SkillGroupArn: The ARN of the skill group to associate the skill to. Required.

Main.Alexa_For_Business.associate_skill_with_users — Method

AssociateSkillWithUsers()

Makes a private skill available for enrolled users to enable on their devices.

Required Parameters

SkillId: The private skill ID you want to make available to enrolled users.

Main.Alexa_For_Business.create_address_book — Method

CreateAddressBook()

Creates an address book with the specified details.

Required Parameters

Name: The name of the address book.

Optional Parameters

ClientRequestToken: A unique, user-specified identifier for the request that ensures idempotency.

Description: The description of the address book.

Main.Alexa_For_Business.create_business_report_schedule — Method

CreateBusinessReportSchedule()

Creates a recurring schedule for usage reports to deliver to the specified S3 location with a specified daily or weekly interval.

Required Parameters

ContentRange: The content range of the reports.

Format: The format of the generated report (individual CSV files or zipped files of individual files).

Optional Parameters

ClientRequestToken: The client request token.

Recurrence: The recurrence of the reports. If this isn't specified, the report will only be delivered one time when the API is called.

S3BucketName: The S3 bucket name of the output reports. If this isn't specified, the report can be retrieved from a download link by calling ListBusinessReportSchedule.

S3KeyPrefix: The S3 key where the report is delivered.

ScheduleName: The name identifier of the schedule.

Tags: The tags for the business report schedule.

Main.Alexa_For_Business.create_conference_provider — Method

CreateConferenceProvider()

Adds a new conference provider under the user's AWS account.

Required Parameters

ConferenceProviderName: The name of the conference provider.

ConferenceProviderType: Represents a type within a list of predefined types.

MeetingSetting: The meeting settings for the conference provider.

Optional Parameters

ClientRequestToken: The request token of the client.

IPDialIn: The IP endpoint and protocol for calling.

PSTNDialIn: The information for PSTN conferencing.

Main.Alexa_For_Business.create_contact — Method

CreateContact()

Creates a contact with the specified details.

Required Parameters

FirstName: The first name of the contact that is used to call the contact on the device.

Optional Parameters

ClientRequestToken: A unique, user-specified identifier for this request that ensures idempotency.

DisplayName: The name of the contact to display on the console.

LastName: The last name of the contact that is used to call the contact on the device.

PhoneNumber: The phone number of the contact in E.164 format. The phone number type defaults to WORK. You can specify PhoneNumber or PhoneNumbers. We recommend that you use PhoneNumbers, which lets you specify the phone number type and multiple numbers.

PhoneNumbers: The list of phone numbers for the contact.

SipAddresses: The list of SIP addresses for the contact.

Main.Alexa_For_Business.create_gateway_group — Method

CreateGatewayGroup()

Creates a gateway group with the specified details.

Required Parameters

ClientRequestToken: A unique, user-specified identifier for the request that ensures idempotency.

Name: The name of the gateway group.

Optional Parameters

Description: The description of the gateway group.

Main.Alexa_For_Business.create_network_profile — Method

CreateNetworkProfile()

Creates a network profile with the specified details.

Required Parameters

ClientRequestToken:

NetworkProfileName: The name of the network profile associated with a device.

SecurityType: The security type of the Wi-Fi network. This can be WPA2ENTERPRISE, WPA2PSK, WPA_PSK, WEP, or OPEN.

Ssid: The SSID of the Wi-Fi network.

Optional Parameters

CertificateAuthorityArn: The ARN of the Private Certificate Authority (PCA) created in AWS Certificate Manager (ACM). This is used to issue certificates to the devices.

CurrentPassword: The current password of the Wi-Fi network.

Description: Detailed information about a device's network profile.

EapMethod: The authentication standard that is used in the EAP framework. Currently, EAP_TLS is supported.

NextPassword: The next, or subsequent, password of the Wi-Fi network. This password is asynchronously transmitted to the device and is used when the password of the network changes to NextPassword.

TrustAnchors: The root certificates of your authentication server that is installed on your devices and used to trust your authentication server during EAP negotiation.

Main.Alexa_For_Business.create_profile — Method

CreateProfile()

Creates a new room profile with the specified details.

Required Parameters

Address: The valid address for the room.

DistanceUnit: The distance unit to be used by devices in the profile.

ProfileName: The name of a room profile.

TemperatureUnit: The temperature unit to be used by devices in the profile.

Timezone: The time zone used by a room profile.

WakeWord: A wake word for Alexa, Echo, Amazon, or a computer.

Optional Parameters

ClientRequestToken: The user-specified token that is used during the creation of a profile.

Locale: The locale of the room profile. (This is currently only available to a limited preview audience.)

MaxVolumeLimit: The maximum volume limit for a room profile.

MeetingRoomConfiguration: The meeting room settings of a room profile.

PSTNEnabled: Whether PSTN calling is enabled.

SetupModeDisabled: Whether room profile setup is enabled.

Tags: The tags for the profile.

Main.Alexa_For_Business.create_room — Method

CreateRoom()

Creates a room with the specified details.

Required Parameters

RoomName: The name for the room.

Optional Parameters

ClientRequestToken: A unique, user-specified identifier for this request that ensures idempotency.

Description: The description for the room.

ProfileArn: The profile ARN for the room. This is required.

ProviderCalendarId: The calendar ARN for the room.

Tags: The tags for the room.

Main.Alexa_For_Business.create_skill_group — Method

CreateSkillGroup()

Creates a skill group with a specified name and description.

Required Parameters

SkillGroupName: The name for the skill group.

Optional Parameters

ClientRequestToken: A unique, user-specified identifier for this request that ensures idempotency.

Description: The description for the skill group.

Tags: The tags for the skill group.

Main.Alexa_For_Business.create_user — Method

CreateUser()

Creates a user.

Required Parameters

UserId: The ARN for the user.

Optional Parameters

ClientRequestToken: A unique, user-specified identifier for this request that ensures idempotency.

Email: The email address for the user.

FirstName: The first name for the user.

LastName: The last name for the user.

Tags: The tags for the user.

Main.Alexa_For_Business.delete_address_book — Method

DeleteAddressBook()

Deletes an address book by the address book ARN.

Required Parameters

AddressBookArn: The ARN of the address book to delete.

Main.Alexa_For_Business.delete_business_report_schedule — Method

DeleteBusinessReportSchedule()

Deletes the recurring report delivery schedule with the specified schedule ARN.

Required Parameters

ScheduleArn: The ARN of the business report schedule.

Main.Alexa_For_Business.delete_conference_provider — Method

DeleteConferenceProvider()

Deletes a conference provider.

Required Parameters

ConferenceProviderArn: The ARN of the conference provider.

Main.Alexa_For_Business.delete_contact — Method

DeleteContact()

Deletes a contact by the contact ARN.

Required Parameters

ContactArn: The ARN of the contact to delete.

Main.Alexa_For_Business.delete_device — Method

DeleteDevice()

Removes a device from Alexa For Business.

Required Parameters

DeviceArn: The ARN of the device for which to request details.

Main.Alexa_For_Business.delete_device_usage_data — Method

DeleteDeviceUsageData()

When this action is called for a specified shared device, it allows authorized users to delete the device's entire previous history of voice input data and associated response data. This action can be called once every 24 hours for a specific shared device.

Required Parameters

DeviceArn: The ARN of the device.

DeviceUsageType: The type of usage data to delete.

Main.Alexa_For_Business.delete_gateway_group — Method

DeleteGatewayGroup()

Deletes a gateway group.

Required Parameters

GatewayGroupArn: The ARN of the gateway group to delete.

Main.Alexa_For_Business.delete_network_profile — Method

DeleteNetworkProfile()

Deletes a network profile by the network profile ARN.

Required Parameters

NetworkProfileArn: The ARN of the network profile associated with a device.

Main.Alexa_For_Business.delete_profile — Method

DeleteProfile()

Deletes a room profile by the profile ARN.

Optional Parameters

ProfileArn: The ARN of the room profile to delete. Required.

Main.Alexa_For_Business.delete_room — Method

DeleteRoom()

Deletes a room by the room ARN.

Optional Parameters

RoomArn: The ARN of the room to delete. Required.

Main.Alexa_For_Business.delete_room_skill_parameter — Method

DeleteRoomSkillParameter()

Deletes room skill parameter details by room, skill, and parameter key ID.

Required Parameters

ParameterKey: The room skill parameter key for which to remove details.

SkillId: The ID of the skill from which to remove the room skill parameter details.

Optional Parameters

RoomArn: The ARN of the room from which to remove the room skill parameter details.

Main.Alexa_For_Business.delete_skill_authorization — Method

DeleteSkillAuthorization()

Unlinks a third-party account from a skill.

Required Parameters

SkillId: The unique identifier of a skill.

Optional Parameters

RoomArn: The room that the skill is authorized for.

Main.Alexa_For_Business.delete_skill_group — Method

DeleteSkillGroup()

Deletes a skill group by skill group ARN.

Optional Parameters

SkillGroupArn: The ARN of the skill group to delete. Required.

Main.Alexa_For_Business.delete_user — Method

DeleteUser()

Deletes a specified user by user ARN and enrollment ARN.

Required Parameters

EnrollmentId: The ARN of the user's enrollment in the organization. Required.

Optional Parameters

UserArn: The ARN of the user to delete in the organization. Required.

Main.Alexa_For_Business.disassociate_contact_from_address_book — Method

DisassociateContactFromAddressBook()

Disassociates a contact from a given address book.

Required Parameters

AddressBookArn: The ARN of the address from which to disassociate the contact.

ContactArn: The ARN of the contact to disassociate from an address book.

Main.Alexa_For_Business.disassociate_device_from_room — Method

DisassociateDeviceFromRoom()

Disassociates a device from its current room. The device continues to be connected to the Wi-Fi network and is still registered to the account. The device settings and skills are removed from the room.

Optional Parameters

DeviceArn: The ARN of the device to disassociate from a room. Required.

Main.Alexa_For_Business.disassociate_skill_from_skill_group — Method

DisassociateSkillFromSkillGroup()

Disassociates a skill from a skill group.

Required Parameters

SkillId: The ARN of a skill group to associate to a skill.

Optional Parameters

SkillGroupArn: The unique identifier of a skill. Required.

Main.Alexa_For_Business.disassociate_skill_from_users — Method

DisassociateSkillFromUsers()

Makes a private skill unavailable for enrolled users and prevents them from enabling it on their devices.

Required Parameters

SkillId: The private skill ID you want to make unavailable for enrolled users.

Main.Alexa_For_Business.disassociate_skill_group_from_room — Method

DisassociateSkillGroupFromRoom()

Disassociates a skill group from a specified room. This disables all skills in the skill group on all devices in the room.

Optional Parameters

RoomArn: The ARN of the room from which the skill group is to be disassociated. Required.

SkillGroupArn: The ARN of the skill group to disassociate from a room. Required.

Main.Alexa_For_Business.forget_smart_home_appliances — Method

ForgetSmartHomeAppliances()

Forgets smart home appliances associated to a room.

Required Parameters

RoomArn: The room that the appliances are associated with.

Main.Alexa_For_Business.get_address_book — Method

GetAddressBook()

Gets address the book details by the address book ARN.

Required Parameters

AddressBookArn: The ARN of the address book for which to request details.

Main.Alexa_For_Business.get_conference_preference — Method

GetConferencePreference()

Retrieves the existing conference preferences.

Main.Alexa_For_Business.get_conference_provider — Method

GetConferenceProvider()

Gets details about a specific conference provider.

Required Parameters

ConferenceProviderArn: The ARN of the newly created conference provider.

Main.Alexa_For_Business.get_contact — Method

GetContact()

Gets the contact details by the contact ARN.

Required Parameters

ContactArn: The ARN of the contact for which to request details.

Main.Alexa_For_Business.get_device — Method

GetDevice()

Gets the details of a device by device ARN.

Optional Parameters

DeviceArn: The ARN of the device for which to request details. Required.

Main.Alexa_For_Business.get_gateway — Method

GetGateway()

Retrieves the details of a gateway.

Required Parameters

GatewayArn: The ARN of the gateway to get.

Main.Alexa_For_Business.get_gateway_group — Method

GetGatewayGroup()

Retrieves the details of a gateway group.

Required Parameters

GatewayGroupArn: The ARN of the gateway group to get.

Main.Alexa_For_Business.get_invitation_configuration — Method

GetInvitationConfiguration()

Retrieves the configured values for the user enrollment invitation email template.

Main.Alexa_For_Business.get_network_profile — Method

GetNetworkProfile()

Gets the network profile details by the network profile ARN.

Required Parameters

NetworkProfileArn: The ARN of the network profile associated with a device.

Main.Alexa_For_Business.get_profile — Method

GetProfile()

Gets the details of a room profile by profile ARN.

Optional Parameters

ProfileArn: The ARN of the room profile for which to request details. Required.

Main.Alexa_For_Business.get_room — Method

GetRoom()

Gets room details by room ARN.

Optional Parameters

RoomArn: The ARN of the room for which to request details. Required.

Main.Alexa_For_Business.get_room_skill_parameter — Method

GetRoomSkillParameter()

Gets room skill parameter details by room, skill, and parameter key ARN.

Required Parameters

ParameterKey: The room skill parameter key for which to get details. Required.

SkillId: The ARN of the skill from which to get the room skill parameter details. Required.

Optional Parameters

RoomArn: The ARN of the room from which to get the room skill parameter details.

Main.Alexa_For_Business.get_skill_group — Method

GetSkillGroup()

Gets skill group details by skill group ARN.

Optional Parameters

SkillGroupArn: The ARN of the skill group for which to get details. Required.

Main.Alexa_For_Business.list_business_report_schedules — Method

ListBusinessReportSchedules()

Lists the details of the schedules that a user configured. A download URL of the report associated with each schedule is returned every time this action is called. A new download URL is returned each time, and is valid for 24 hours.

Optional Parameters

MaxResults: The maximum number of schedules listed in the call.

NextToken: The token used to list the remaining schedules from the previous API call.

Main.Alexa_For_Business.list_conference_providers — Method

ListConferenceProviders()

Lists conference providers under a specific AWS account.

Optional Parameters

MaxResults: The maximum number of conference providers to be returned, per paginated calls.

NextToken: The tokens used for pagination.

Main.Alexa_For_Business.list_device_events — Method

ListDeviceEvents()

Lists the device event history, including device connection status, for up to 30 days.

Required Parameters

DeviceArn: The ARN of a device.

Optional Parameters

EventType: The event type to filter device events. If EventType isn't specified, this returns a list of all device events in reverse chronological order. If EventType is specified, this returns a list of device events for that EventType in reverse chronological order.

MaxResults: The maximum number of results to include in the response. The default value is 50. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

NextToken: An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response only includes results beyond the token, up to the value specified by MaxResults. When the end of results is reached, the response has a value of null.

Main.Alexa_For_Business.list_gateway_groups — Method

ListGatewayGroups()

Retrieves a list of gateway group summaries. Use GetGatewayGroup to retrieve details of a specific gateway group.

Optional Parameters

MaxResults: The maximum number of gateway group summaries to return. The default is 50.

NextToken: The token used to paginate though multiple pages of gateway group summaries.

Main.Alexa_For_Business.list_gateways — Method

ListGateways()

Retrieves a list of gateway summaries. Use GetGateway to retrieve details of a specific gateway. An optional gateway group ARN can be provided to only retrieve gateway summaries of gateways that are associated with that gateway group ARN.

Optional Parameters

GatewayGroupArn: The gateway group ARN for which to list gateways.

MaxResults: The maximum number of gateway summaries to return. The default is 50.

NextToken: The token used to paginate though multiple pages of gateway summaries.

Main.Alexa_For_Business.list_skills — Method

ListSkills()

Lists all enabled skills in a specific skill group.

Optional Parameters

EnablementType: Whether the skill is enabled under the user's account.

MaxResults: The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

NextToken: An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

SkillGroupArn: The ARN of the skill group for which to list enabled skills.

SkillType: Whether the skill is publicly available or is a private skill.

Main.Alexa_For_Business.list_skills_store_categories — Method

ListSkillsStoreCategories()

Lists all categories in the Alexa skill store.

Optional Parameters

MaxResults: The maximum number of categories returned, per paginated calls.

NextToken: The tokens used for pagination.

Main.Alexa_For_Business.list_skills_store_skills_by_category — Method

ListSkillsStoreSkillsByCategory()

Lists all skills in the Alexa skill store by category.

Required Parameters

CategoryId: The category ID for which the skills are being retrieved from the skill store.

Optional Parameters

MaxResults: The maximum number of skills returned per paginated calls.

NextToken: The tokens used for pagination.

Main.Alexa_For_Business.list_smart_home_appliances — Method

ListSmartHomeAppliances()

Lists all of the smart home appliances associated with a room.

Required Parameters

RoomArn: The room that the appliances are associated with.

Optional Parameters

MaxResults: The maximum number of appliances to be returned, per paginated calls.

NextToken: The tokens used for pagination.

Main.Alexa_For_Business.list_tags — Method

ListTags()

Lists all tags for the specified resource.

Required Parameters

Arn: The ARN of the specified resource for which to list tags.

Optional Parameters

MaxResults: The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

NextToken: An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

Main.Alexa_For_Business.put_conference_preference — Method

PutConferencePreference()

Sets the conference preferences on a specific conference provider at the account level.

Required Parameters

ConferencePreference: The conference preference of a specific conference provider.

Main.Alexa_For_Business.put_invitation_configuration — Method

PutInvitationConfiguration()

Configures the email template for the user enrollment invitation with the specified attributes.

Required Parameters

OrganizationName: The name of the organization sending the enrollment invite to a user.

Optional Parameters

ContactEmail: The email ID of the organization or individual contact that the enrolled user can use.

PrivateSkillIds: The list of private skill IDs that you want to recommend to the user to enable in the invitation.

Main.Alexa_For_Business.put_room_skill_parameter — Method

PutRoomSkillParameter()

Updates room skill parameter details by room, skill, and parameter key ID. Not all skills have a room skill parameter.

Required Parameters

RoomSkillParameter: The updated room skill parameter. Required.

SkillId: The ARN of the skill associated with the room skill parameter. Required.

Optional Parameters

RoomArn: The ARN of the room associated with the room skill parameter. Required.

Main.Alexa_For_Business.put_skill_authorization — Method

PutSkillAuthorization()

Links a user's account to a third-party skill provider. If this API operation is called by an assumed IAM role, the skill being linked must be a private skill. Also, the skill must be owned by the AWS account that assumed the IAM role.

Required Parameters

AuthorizationResult: The authorization result specific to OAUTH code grant output. "Code” must be populated in the AuthorizationResult map to establish the authorization.

SkillId: The unique identifier of a skill.

Optional Parameters

RoomArn: The room that the skill is authorized for.

Main.Alexa_For_Business.register_avsdevice — Method

RegisterAVSDevice()

Registers an Alexa-enabled device built by an Original Equipment Manufacturer (OEM) using Alexa Voice Service (AVS).

Required Parameters

AmazonId: The device type ID for your AVS device generated by Amazon when the OEM creates a new product on Amazon's Developer Console.

ClientId: The client ID of the OEM used for code-based linking authorization on an AVS device.

ProductId: The product ID used to identify your AVS device during authorization.

UserCode: The code that is obtained after your AVS device has made a POST request to LWA as a part of the Device Authorization Request component of the OAuth code-based linking specification.

Optional Parameters

DeviceSerialNumber: The key generated by the OEM that uniquely identifies a specified instance of your AVS device.

RoomArn: The ARN of the room with which to associate your AVS device.

Main.Alexa_For_Business.reject_skill — Method

RejectSkill()

Disassociates a skill from the organization under a user's AWS account. If the skill is a private skill, it moves to an AcceptStatus of PENDING. Any private or public skill that is rejected can be added later by calling the ApproveSkill API.

Required Parameters

SkillId: The unique identifier of the skill.

Main.Alexa_For_Business.resolve_room — Method

ResolveRoom()

Determines the details for the room from which a skill request was invoked. This operation is used by skill developers.

Required Parameters

SkillId: The ARN of the skill that was requested. Required.

UserId: The ARN of the user. Required.

Main.Alexa_For_Business.revoke_invitation — Method

RevokeInvitation()

Revokes an invitation and invalidates the enrollment URL.

Optional Parameters

EnrollmentId: The ARN of the enrollment invitation to revoke. Required.

UserArn: The ARN of the user for whom to revoke an enrollment invitation. Required.

Main.Alexa_For_Business.search_address_books — Method

SearchAddressBooks()

Searches address books and lists the ones that meet a set of filter and sort criteria.

Optional Parameters

Filters: The filters to use to list a specified set of address books. The supported filter key is AddressBookName.

MaxResults: The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

NextToken: An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response only includes results beyond the token, up to the value specified by MaxResults.

SortCriteria: The sort order to use in listing the specified set of address books. The supported sort key is AddressBookName.

Main.Alexa_For_Business.search_contacts — Method

SearchContacts()

Searches contacts and lists the ones that meet a set of filter and sort criteria.

Optional Parameters

Filters: The filters to use to list a specified set of address books. The supported filter keys are DisplayName, FirstName, LastName, and AddressBookArns.

MaxResults: The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

NextToken: An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response only includes results beyond the token, up to the value specified by MaxResults.

SortCriteria: The sort order to use in listing the specified set of contacts. The supported sort keys are DisplayName, FirstName, and LastName.

Main.Alexa_For_Business.search_devices — Method

SearchDevices()

Searches devices and lists the ones that meet a set of filter criteria.

Optional Parameters

Filters: The filters to use to list a specified set of devices. Supported filter keys are DeviceName, DeviceStatus, DeviceStatusDetailCode, RoomName, DeviceType, DeviceSerialNumber, UnassociatedOnly, ConnectionStatus (ONLINE and OFFLINE), NetworkProfileName, NetworkProfileArn, Feature, and FailureCode.

MaxResults: The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

NextToken: An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

SortCriteria: The sort order to use in listing the specified set of devices. Supported sort keys are DeviceName, DeviceStatus, RoomName, DeviceType, DeviceSerialNumber, ConnectionStatus, NetworkProfileName, NetworkProfileArn, Feature, and FailureCode.

Main.Alexa_For_Business.search_network_profiles — Method

SearchNetworkProfiles()

Searches network profiles and lists the ones that meet a set of filter and sort criteria.

Optional Parameters

Filters: The filters to use to list a specified set of network profiles. Valid filters are NetworkProfileName, Ssid, and SecurityType.

MaxResults: The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

NextToken: An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

SortCriteria: The sort order to use to list the specified set of network profiles. Valid sort criteria includes NetworkProfileName, Ssid, and SecurityType.

Main.Alexa_For_Business.search_profiles — Method

SearchProfiles()

Searches room profiles and lists the ones that meet a set of filter criteria.

Optional Parameters

Filters: The filters to use to list a specified set of room profiles. Supported filter keys are ProfileName and Address. Required.

MaxResults: The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

NextToken: An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

SortCriteria: The sort order to use in listing the specified set of room profiles. Supported sort keys are ProfileName and Address.

Main.Alexa_For_Business.search_rooms — Method

SearchRooms()

Searches rooms and lists the ones that meet a set of filter and sort criteria.

Optional Parameters

Filters: The filters to use to list a specified set of rooms. The supported filter keys are RoomName and ProfileName.

MaxResults: The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

NextToken: An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

SortCriteria: The sort order to use in listing the specified set of rooms. The supported sort keys are RoomName and ProfileName.

Main.Alexa_For_Business.search_skill_groups — Method

SearchSkillGroups()

Searches skill groups and lists the ones that meet a set of filter and sort criteria.

Optional Parameters

Filters: The filters to use to list a specified set of skill groups. The supported filter key is SkillGroupName.

MaxResults: The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

NextToken: An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults. Required.

SortCriteria: The sort order to use in listing the specified set of skill groups. The supported sort key is SkillGroupName.

Main.Alexa_For_Business.search_users — Method

SearchUsers()

Searches users and lists the ones that meet a set of filter and sort criteria.

Optional Parameters

Filters: The filters to use for listing a specific set of users. Required. Supported filter keys are UserId, FirstName, LastName, Email, and EnrollmentStatus.

MaxResults: The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved. Required.

NextToken: An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults. Required.

SortCriteria: The sort order to use in listing the filtered set of users. Required. Supported sort keys are UserId, FirstName, LastName, Email, and EnrollmentStatus.

Main.Alexa_For_Business.send_announcement — Method

SendAnnouncement()

Triggers an asynchronous flow to send text, SSML, or audio announcements to rooms that are identified by a search or filter.

Required Parameters

ClientRequestToken: The unique, user-specified identifier for the request that ensures idempotency.

Content: The announcement content. This can contain only one of the three possible announcement types (text, SSML or audio).

RoomFilters: The filters to use to send an announcement to a specified list of rooms. The supported filter keys are RoomName, ProfileName, RoomArn, and ProfileArn. To send to all rooms, specify an empty RoomFilters list.

Optional Parameters

TimeToLiveInSeconds: The time to live for an announcement. Default is 300. If delivery doesn't occur within this time, the announcement is not delivered.

Main.Alexa_For_Business.send_invitation — Method

SendInvitation()

Sends an enrollment invitation email with a URL to a user. The URL is valid for 30 days or until you call this operation again, whichever comes first.

Optional Parameters

UserArn: The ARN of the user to whom to send an invitation. Required.

Main.Alexa_For_Business.start_device_sync — Method

StartDeviceSync()

Resets a device and its account to the known default settings. This clears all information and settings set by previous users in the following ways: Bluetooth - This unpairs all bluetooth devices paired with your echo device. Volume - This resets the echo device's volume to the default value. Notifications - This clears all notifications from your echo device. Lists - This clears all to-do items from your echo device. Settings - This internally syncs the room's profile (if the device is assigned to a room), contacts, address books, delegation access for account linking, and communications (if enabled on the room profile).

Required Parameters

Features: Request structure to start the device sync. Required.

Optional Parameters

DeviceArn: The ARN of the device to sync. Required.

RoomArn: The ARN of the room with which the device to sync is associated. Required.

Main.Alexa_For_Business.start_smart_home_appliance_discovery — Method

StartSmartHomeApplianceDiscovery()

Initiates the discovery of any smart home appliances associated with the room.

Required Parameters

RoomArn: The room where smart home appliance discovery was initiated.

Main.Alexa_For_Business.tag_resource — Method

TagResource()

Adds metadata tags to a specified resource.

Required Parameters

Arn: The ARN of the resource to which to add metadata tags. Required.

Tags: The tags to be added to the specified resource. Do not provide system tags. Required.

Main.Alexa_For_Business.untag_resource — Method

UntagResource()

Removes metadata tags from a specified resource.

Required Parameters

Arn: The ARN of the resource from which to remove metadata tags. Required.

TagKeys: The tags to be removed from the specified resource. Do not provide system tags. Required.

Main.Alexa_For_Business.update_address_book — Method

UpdateAddressBook()

Updates address book details by the address book ARN.

Required Parameters

AddressBookArn: The ARN of the room to update.

Optional Parameters

Description: The updated description of the room.

Name: The updated name of the room.

Main.Alexa_For_Business.update_business_report_schedule — Method

UpdateBusinessReportSchedule()

Updates the configuration of the report delivery schedule with the specified schedule ARN.

Required Parameters

ScheduleArn: The ARN of the business report schedule.

Optional Parameters

Format: The format of the generated report (individual CSV files or zipped files of individual files).

Recurrence: The recurrence of the reports.

S3BucketName: The S3 location of the output reports.

S3KeyPrefix: The S3 key where the report is delivered.

ScheduleName: The name identifier of the schedule.

Main.Alexa_For_Business.update_conference_provider — Method

UpdateConferenceProvider()

Updates an existing conference provider's settings.

Required Parameters

ConferenceProviderArn: The ARN of the conference provider.

ConferenceProviderType: The type of the conference provider.

MeetingSetting: The meeting settings for the conference provider.

Optional Parameters

IPDialIn: The IP endpoint and protocol for calling.

PSTNDialIn: The information for PSTN conferencing.

Main.Alexa_For_Business.update_contact — Method

UpdateContact()

Updates the contact details by the contact ARN.

Required Parameters

ContactArn: The ARN of the contact to update.

Optional Parameters

DisplayName: The updated display name of the contact.

FirstName: The updated first name of the contact.

LastName: The updated last name of the contact.

PhoneNumber: The updated phone number of the contact. The phone number type defaults to WORK. You can either specify PhoneNumber or PhoneNumbers. We recommend that you use PhoneNumbers, which lets you specify the phone number type and multiple numbers.

PhoneNumbers: The list of phone numbers for the contact.

SipAddresses: The list of SIP addresses for the contact.

Main.Alexa_For_Business.update_device — Method

UpdateDevice()

Updates the device name by device ARN.

Optional Parameters

DeviceArn: The ARN of the device to update. Required.

DeviceName: The updated device name. Required.

Main.Alexa_For_Business.update_gateway — Method

UpdateGateway()

Updates the details of a gateway. If any optional field is not provided, the existing corresponding value is left unmodified.

Required Parameters

GatewayArn: The ARN of the gateway to update.

Optional Parameters

Description: The updated description of the gateway.

Name: The updated name of the gateway.

SoftwareVersion: The updated software version of the gateway. The gateway automatically updates its software version during normal operation.

Main.Alexa_For_Business.update_gateway_group — Method

UpdateGatewayGroup()

Updates the details of a gateway group. If any optional field is not provided, the existing corresponding value is left unmodified.

Required Parameters

GatewayGroupArn: The ARN of the gateway group to update.

Optional Parameters

Description: The updated description of the gateway group.

Name: The updated name of the gateway group.

Main.Alexa_For_Business.update_network_profile — Method

UpdateNetworkProfile()

Updates a network profile by the network profile ARN.

Required Parameters

NetworkProfileArn: The ARN of the network profile associated with a device.

Optional Parameters

CertificateAuthorityArn: The ARN of the Private Certificate Authority (PCA) created in AWS Certificate Manager (ACM). This is used to issue certificates to the devices.

CurrentPassword: The current password of the Wi-Fi network.

Description: Detailed information about a device's network profile.

NetworkProfileName: The name of the network profile associated with a device.

NextPassword: The next, or subsequent, password of the Wi-Fi network. This password is asynchronously transmitted to the device and is used when the password of the network changes to NextPassword.

TrustAnchors: The root certificate(s) of your authentication server that will be installed on your devices and used to trust your authentication server during EAP negotiation.

Main.Alexa_For_Business.update_profile — Method

UpdateProfile()

Updates an existing room profile by room profile ARN.

Optional Parameters

Address: The updated address for the room profile.

DistanceUnit: The updated distance unit for the room profile.

IsDefault: Sets the profile as default if selected. If this is missing, no update is done to the default status.

Locale: The updated locale for the room profile. (This is currently only available to a limited preview audience.)

MaxVolumeLimit: The updated maximum volume limit for the room profile.

MeetingRoomConfiguration: The updated meeting room settings of a room profile.

PSTNEnabled: Whether the PSTN setting of the room profile is enabled.

ProfileArn: The ARN of the room profile to update. Required.

ProfileName: The updated name for the room profile.

SetupModeDisabled: Whether the setup mode of the profile is enabled.

TemperatureUnit: The updated temperature unit for the room profile.

Timezone: The updated timezone for the room profile.

WakeWord: The updated wake word for the room profile.

Main.Alexa_For_Business.update_room — Method

UpdateRoom()

Updates room details by room ARN.

Optional Parameters

Description: The updated description for the room.

ProfileArn: The updated profile ARN for the room.

ProviderCalendarId: The updated provider calendar ARN for the room.

RoomArn: The ARN of the room to update.

RoomName: The updated name for the room.

Main.Alexa_For_Business.update_skill_group — Method

UpdateSkillGroup()

Updates skill group details by skill group ARN.

Optional Parameters

Description: The updated description for the skill group.

SkillGroupArn: The ARN of the skill group to update.

SkillGroupName: The updated name for the skill group.