Amplify

This page documents function available when using the Amplify module, created with @service Amplify.

Index

Main.Amplify.create_app

Main.Amplify.create_backend_environment

Main.Amplify.create_branch

Main.Amplify.create_deployment

Main.Amplify.create_domain_association

Main.Amplify.create_webhook

Main.Amplify.delete_app

Main.Amplify.delete_backend_environment

Main.Amplify.delete_branch

Main.Amplify.delete_domain_association

Main.Amplify.delete_job

Main.Amplify.delete_webhook

Main.Amplify.generate_access_logs

Main.Amplify.get_app

Main.Amplify.get_artifact_url

Main.Amplify.get_backend_environment

Main.Amplify.get_branch

Main.Amplify.get_domain_association

Main.Amplify.get_job

Main.Amplify.get_webhook

Main.Amplify.list_apps

Main.Amplify.list_artifacts

Main.Amplify.list_backend_environments

Main.Amplify.list_branches

Main.Amplify.list_domain_associations

Main.Amplify.list_jobs

Main.Amplify.list_tags_for_resource

Main.Amplify.list_webhooks

Main.Amplify.start_deployment

Main.Amplify.start_job

Main.Amplify.stop_job

Main.Amplify.tag_resource

Main.Amplify.untag_resource

Main.Amplify.update_app

Main.Amplify.update_branch

Main.Amplify.update_domain_association

Main.Amplify.update_webhook

Documentation

Main.Amplify.create_app — Method

create_app(name)
create_app(name, params::Dict{String,<:Any})

Creates a new Amplify app.

Arguments

name: The name for an Amplify app.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"accessToken": The personal access token for a third-party source control system for an Amplify app. The personal access token is used to create a webhook and a read-only deploy key. The token is not stored.

"autoBranchCreationConfig": The automated branch creation configuration for an Amplify app.

"autoBranchCreationPatterns": The automated branch creation glob patterns for an Amplify app.

"basicAuthCredentials": The credentials for basic authorization for an Amplify app.

"buildSpec": The build specification (build spec) for an Amplify app.

"customHeaders": The custom HTTP headers for an Amplify app.

"customRules": The custom rewrite and redirect rules for an Amplify app.

"description": The description for an Amplify app.

"enableAutoBranchCreation": Enables automated branch creation for an Amplify app.

"enableBasicAuth": Enables basic authorization for an Amplify app. This will apply to all branches that are part of this app.

"enableBranchAutoBuild": Enables the auto building of branches for an Amplify app.

"enableBranchAutoDeletion": Automatically disconnects a branch in the Amplify Console when you delete a branch from your Git repository.

"environmentVariables": The environment variables map for an Amplify app.

"iamServiceRoleArn": The AWS Identity and Access Management (IAM) service role for an Amplify app.

"oauthToken": The OAuth token for a third-party source control system for an Amplify app. The OAuth token is used to create a webhook and a read-only deploy key. The OAuth token is not stored.

"platform": The platform or framework for an Amplify app.

"repository": The repository for an Amplify app.

"tags": The tag for an Amplify app.

Main.Amplify.create_backend_environment — Method

create_backend_environment(app_id, environment_name)
create_backend_environment(app_id, environment_name, params::Dict{String,<:Any})

Creates a new backend environment for an Amplify app.

Arguments

app_id: The unique ID for an Amplify app.

environment_name: The name for the backend environment.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"deploymentArtifacts": The name of deployment artifacts.

"stackName": The AWS CloudFormation stack name of a backend environment.

Main.Amplify.create_branch — Method

create_branch(app_id, branch_name)
create_branch(app_id, branch_name, params::Dict{String,<:Any})

Creates a new branch for an Amplify app.

Arguments

app_id: The unique ID for an Amplify app.

branch_name: The name for the branch.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"backendEnvironmentArn": The Amazon Resource Name (ARN) for a backend environment that is part of an Amplify app.

"basicAuthCredentials": The basic authorization credentials for the branch.

"buildSpec": The build specification (build spec) for the branch.

"description": The description for the branch.

"displayName": The display name for a branch. This is used as the default domain prefix.

"enableAutoBuild": Enables auto building for the branch.

"enableBasicAuth": Enables basic authorization for the branch.

"enableNotification": Enables notifications for the branch.

"enablePerformanceMode": Enables performance mode for the branch. Performance mode optimizes for faster hosting performance by keeping content cached at the edge for a longer interval. When performance mode is enabled, hosting configuration or code changes can take up to 10 minutes to roll out.

"enablePullRequestPreview": Enables pull request previews for this branch.

"environmentVariables": The environment variables for the branch.

"framework": The framework for the branch.

"pullRequestEnvironmentName": The Amplify environment name for the pull request.

"stage": Describes the current stage for the branch.

"tags": The tag for the branch.

"ttl": The content Time To Live (TTL) for the website in seconds.

Main.Amplify.create_deployment — Method

create_deployment(app_id, branch_name)
create_deployment(app_id, branch_name, params::Dict{String,<:Any})

Creates a deployment for a manually deployed Amplify app. Manually deployed apps are not connected to a repository.

Arguments

app_id: The unique ID for an Amplify app.

branch_name: The name for the branch, for the job.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"fileMap": An optional file map that contains the file name as the key and the file content md5 hash as the value. If this argument is provided, the service will generate a unique upload URL per file. Otherwise, the service will only generate a single upload URL for the zipped files.

Main.Amplify.create_domain_association — Method

create_domain_association(app_id, domain_name, sub_domain_settings)
create_domain_association(app_id, domain_name, sub_domain_settings, params::Dict{String,<:Any})

Creates a new domain association for an Amplify app. This action associates a custom domain with the Amplify app

Arguments

app_id: The unique ID for an Amplify app.

domain_name: The domain name for the domain association.

sub_domain_settings: The setting for the subdomain.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"autoSubDomainCreationPatterns": Sets the branch patterns for automatic subdomain creation.

"autoSubDomainIAMRole": The required AWS Identity and Access Management (IAM) service role for the Amazon Resource Name (ARN) for automatically creating subdomains.

"enableAutoSubDomain": Enables the automated creation of subdomains for branches.

Main.Amplify.create_webhook — Method

create_webhook(app_id, branch_name)
create_webhook(app_id, branch_name, params::Dict{String,<:Any})

Creates a new webhook on an Amplify app.

Arguments

app_id: The unique ID for an Amplify app.

branch_name: The name for a branch that is part of an Amplify app.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The description for a webhook.

Main.Amplify.delete_app — Method

delete_app(app_id)
delete_app(app_id, params::Dict{String,<:Any})

Deletes an existing Amplify app specified by an app ID.

Arguments

app_id: The unique ID for an Amplify app.

Main.Amplify.delete_backend_environment — Method

delete_backend_environment(app_id, environment_name)
delete_backend_environment(app_id, environment_name, params::Dict{String,<:Any})

Deletes a backend environment for an Amplify app.

Arguments

app_id: The unique ID of an Amplify app.

environment_name: The name of a backend environment of an Amplify app.

Main.Amplify.delete_branch — Method

delete_branch(app_id, branch_name)
delete_branch(app_id, branch_name, params::Dict{String,<:Any})

Deletes a branch for an Amplify app.

Arguments

app_id: The unique ID for an Amplify app.

branch_name: The name for the branch.

Main.Amplify.delete_domain_association — Method

delete_domain_association(app_id, domain_name)
delete_domain_association(app_id, domain_name, params::Dict{String,<:Any})

Deletes a domain association for an Amplify app.

Arguments

app_id: The unique id for an Amplify app.

domain_name: The name of the domain.

Main.Amplify.delete_job — Method

delete_job(app_id, branch_name, job_id)
delete_job(app_id, branch_name, job_id, params::Dict{String,<:Any})

Deletes a job for a branch of an Amplify app.

Arguments

app_id: The unique ID for an Amplify app.

branch_name: The name for the branch, for the job.

job_id: The unique ID for the job.

Main.Amplify.delete_webhook — Method

delete_webhook(webhook_id)
delete_webhook(webhook_id, params::Dict{String,<:Any})

Deletes a webhook.

Arguments

webhook_id: The unique ID for a webhook.

Main.Amplify.generate_access_logs — Method

generate_access_logs(app_id, domain_name)
generate_access_logs(app_id, domain_name, params::Dict{String,<:Any})

Returns the website access logs for a specific time range using a presigned URL.

Arguments

app_id: The unique ID for an Amplify app.

domain_name: The name of the domain.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"endTime": The time at which the logs should end. The time range specified is inclusive of the end time.

"startTime": The time at which the logs should start. The time range specified is inclusive of the start time.

Main.Amplify.get_app — Method

get_app(app_id)
get_app(app_id, params::Dict{String,<:Any})

Returns an existing Amplify app by appID.

Arguments

app_id: The unique ID for an Amplify app.

Main.Amplify.get_artifact_url — Method

get_artifact_url(artifact_id)
get_artifact_url(artifact_id, params::Dict{String,<:Any})

Returns the artifact info that corresponds to an artifact id.

Arguments

artifact_id: The unique ID for an artifact.

Main.Amplify.get_backend_environment — Method

get_backend_environment(app_id, environment_name)
get_backend_environment(app_id, environment_name, params::Dict{String,<:Any})

Returns a backend environment for an Amplify app.

Arguments

app_id: The unique id for an Amplify app.

environment_name: The name for the backend environment.

Main.Amplify.get_branch — Method

get_branch(app_id, branch_name)
get_branch(app_id, branch_name, params::Dict{String,<:Any})

Returns a branch for an Amplify app.

Arguments

app_id: The unique ID for an Amplify app.

branch_name: The name for the branch.

Main.Amplify.get_domain_association — Method

get_domain_association(app_id, domain_name)
get_domain_association(app_id, domain_name, params::Dict{String,<:Any})

Returns the domain information for an Amplify app.

Arguments

app_id: The unique id for an Amplify app.

domain_name: The name of the domain.

Main.Amplify.get_job — Method

get_job(app_id, branch_name, job_id)
get_job(app_id, branch_name, job_id, params::Dict{String,<:Any})

Returns a job for a branch of an Amplify app.

Arguments

app_id: The unique ID for an Amplify app.

branch_name: The branch name for the job.

job_id: The unique ID for the job.

Main.Amplify.get_webhook — Method

get_webhook(webhook_id)
get_webhook(webhook_id, params::Dict{String,<:Any})

Returns the webhook information that corresponds to a specified webhook ID.

Arguments

webhook_id: The unique ID for a webhook.

Main.Amplify.list_apps — Method

list_apps()
list_apps(params::Dict{String,<:Any})

Returns a list of the existing Amplify apps.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of records to list in a single response.

"nextToken": A pagination token. If non-null, the pagination token is returned in a result. Pass its value in another request to retrieve more entries.

Main.Amplify.list_artifacts — Method

list_artifacts(app_id, branch_name, job_id)
list_artifacts(app_id, branch_name, job_id, params::Dict{String,<:Any})

Returns a list of artifacts for a specified app, branch, and job.

Arguments

app_id: The unique ID for an Amplify app.

branch_name: The name of a branch that is part of an Amplify app.

job_id: The unique ID for a job.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of records to list in a single response.

"nextToken": A pagination token. Set to null to start listing artifacts from start. If a non-null pagination token is returned in a result, pass its value in here to list more artifacts.

Main.Amplify.list_backend_environments — Method

list_backend_environments(app_id)
list_backend_environments(app_id, params::Dict{String,<:Any})

Lists the backend environments for an Amplify app.

Arguments

app_id: The unique ID for an Amplify app.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"environmentName": The name of the backend environment

"maxResults": The maximum number of records to list in a single response.

"nextToken": A pagination token. Set to null to start listing backend environments from the start. If a non-null pagination token is returned in a result, pass its value in here to list more backend environments.

Main.Amplify.list_branches — Method

list_branches(app_id)
list_branches(app_id, params::Dict{String,<:Any})

Lists the branches of an Amplify app.

Arguments

app_id: The unique ID for an Amplify app.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of records to list in a single response.

"nextToken": A pagination token. Set to null to start listing branches from the start. If a non-null pagination token is returned in a result, pass its value in here to list more branches.

Main.Amplify.list_domain_associations — Method

list_domain_associations(app_id)
list_domain_associations(app_id, params::Dict{String,<:Any})

Returns the domain associations for an Amplify app.

Arguments

app_id: The unique ID for an Amplify app.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of records to list in a single response.

"nextToken": A pagination token. Set to null to start listing apps from the start. If non-null, a pagination token is returned in a result. Pass its value in here to list more projects.

Main.Amplify.list_jobs — Method

list_jobs(app_id, branch_name)
list_jobs(app_id, branch_name, params::Dict{String,<:Any})

Lists the jobs for a branch of an Amplify app.

Arguments

app_id: The unique ID for an Amplify app.

branch_name: The name for a branch.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of records to list in a single response.

"nextToken": A pagination token. Set to null to start listing steps from the start. If a non-null pagination token is returned in a result, pass its value in here to list more steps.

Main.Amplify.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Returns a list of tags for a specified Amazon Resource Name (ARN).

Arguments

resource_arn: The Amazon Resource Name (ARN) to use to list tags.

Main.Amplify.list_webhooks — Method

list_webhooks(app_id)
list_webhooks(app_id, params::Dict{String,<:Any})

Returns a list of webhooks for an Amplify app.

Arguments

app_id: The unique ID for an Amplify app.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of records to list in a single response.

"nextToken": A pagination token. Set to null to start listing webhooks from the start. If non-null,the pagination token is returned in a result. Pass its value in here to list more webhooks.

Main.Amplify.start_deployment — Method

start_deployment(app_id, branch_name)
start_deployment(app_id, branch_name, params::Dict{String,<:Any})

Starts a deployment for a manually deployed app. Manually deployed apps are not connected to a repository.

Arguments

app_id: The unique ID for an Amplify app.

branch_name: The name for the branch, for the job.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"jobId": The job ID for this deployment, generated by the create deployment request.

"sourceUrl": The source URL for this deployment, used when calling start deployment without create deployment. The source URL can be any HTTP GET URL that is publicly accessible and downloads a single .zip file.

Main.Amplify.start_job — Method

start_job(app_id, branch_name, job_type)
start_job(app_id, branch_name, job_type, params::Dict{String,<:Any})

Starts a new job for a branch of an Amplify app.

Arguments

app_id: The unique ID for an Amplify app.

branch_name: The branch name for the job.

job_type: Describes the type for the job. The job type RELEASE starts a new job with the latest change from the specified branch. This value is available only for apps that are connected to a repository. The job type RETRY retries an existing job. If the job type value is RETRY, the jobId is also required.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"commitId": The commit ID from a third-party repository provider for the job.

"commitMessage": The commit message from a third-party repository provider for the job.

"commitTime": The commit date and time for the job.

"jobId": The unique ID for an existing job. This is required if the value of jobType is RETRY.

"jobReason": A descriptive reason for starting this job.

Main.Amplify.stop_job — Method

stop_job(app_id, branch_name, job_id)
stop_job(app_id, branch_name, job_id, params::Dict{String,<:Any})

Stops a job that is in progress for a branch of an Amplify app.

Arguments

app_id: The unique ID for an Amplify app.

branch_name: The name for the branch, for the job.

job_id: The unique id for the job.

Main.Amplify.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Tags the resource with a tag key and value.

Arguments

resource_arn: The Amazon Resource Name (ARN) to use to tag a resource.

tags: The tags used to tag the resource.

Main.Amplify.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Untags a resource with a specified Amazon Resource Name (ARN).

Arguments

resource_arn: The Amazon Resource Name (ARN) to use to untag a resource.

tag_keys: The tag keys to use to untag a resource.

Main.Amplify.update_app — Method

update_app(app_id)
update_app(app_id, params::Dict{String,<:Any})

Updates an existing Amplify app.

Arguments

app_id: The unique ID for an Amplify app.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"accessToken": The personal access token for a third-party source control system for an Amplify app. The token is used to create webhook and a read-only deploy key. The token is not stored.

"autoBranchCreationConfig": The automated branch creation configuration for an Amplify app.

"autoBranchCreationPatterns": Describes the automated branch creation glob patterns for an Amplify app.

"basicAuthCredentials": The basic authorization credentials for an Amplify app.

"buildSpec": The build specification (build spec) for an Amplify app.

"customHeaders": The custom HTTP headers for an Amplify app.

"customRules": The custom redirect and rewrite rules for an Amplify app.

"description": The description for an Amplify app.

"enableAutoBranchCreation": Enables automated branch creation for an Amplify app.

"enableBasicAuth": Enables basic authorization for an Amplify app.

"enableBranchAutoBuild": Enables branch auto-building for an Amplify app.

"enableBranchAutoDeletion": Automatically disconnects a branch in the Amplify Console when you delete a branch from your Git repository.

"environmentVariables": The environment variables for an Amplify app.

"iamServiceRoleArn": The AWS Identity and Access Management (IAM) service role for an Amplify app.

"name": The name for an Amplify app.

"oauthToken": The OAuth token for a third-party source control system for an Amplify app. The token is used to create a webhook and a read-only deploy key. The OAuth token is not stored.

"platform": The platform for an Amplify app.

"repository": The name of the repository for an Amplify app

Main.Amplify.update_branch — Method

update_branch(app_id, branch_name)
update_branch(app_id, branch_name, params::Dict{String,<:Any})

Updates a branch for an Amplify app.

Arguments

app_id: The unique ID for an Amplify app.

branch_name: The name for the branch.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"backendEnvironmentArn": The Amazon Resource Name (ARN) for a backend environment that is part of an Amplify app.

"basicAuthCredentials": The basic authorization credentials for the branch.

"buildSpec": The build specification (build spec) for the branch.

"description": The description for the branch.

"displayName": The display name for a branch. This is used as the default domain prefix.

"enableAutoBuild": Enables auto building for the branch.

"enableBasicAuth": Enables basic authorization for the branch.

"enableNotification": Enables notifications for the branch.

"enablePerformanceMode": Enables performance mode for the branch. Performance mode optimizes for faster hosting performance by keeping content cached at the edge for a longer interval. When performance mode is enabled, hosting configuration or code changes can take up to 10 minutes to roll out.

"enablePullRequestPreview": Enables pull request previews for this branch.

"environmentVariables": The environment variables for the branch.

"framework": The framework for the branch.

"pullRequestEnvironmentName": The Amplify environment name for the pull request.

"stage": Describes the current stage for the branch.

"ttl": The content Time to Live (TTL) for the website in seconds.

Main.Amplify.update_domain_association — Method

update_domain_association(app_id, domain_name, sub_domain_settings)
update_domain_association(app_id, domain_name, sub_domain_settings, params::Dict{String,<:Any})

Creates a new domain association for an Amplify app.

Arguments

app_id: The unique ID for an Amplify app.

domain_name: The name of the domain.

sub_domain_settings: Describes the settings for the subdomain.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"autoSubDomainCreationPatterns": Sets the branch patterns for automatic subdomain creation.

"autoSubDomainIAMRole": The required AWS Identity and Access Management (IAM) service role for the Amazon Resource Name (ARN) for automatically creating subdomains.

"enableAutoSubDomain": Enables the automated creation of subdomains for branches.

Main.Amplify.update_webhook — Method

update_webhook(webhook_id)
update_webhook(webhook_id, params::Dict{String,<:Any})

Updates a webhook.

Arguments

webhook_id: The unique ID for a webhook.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"branchName": The name for a branch that is part of an Amplify app.

"description": The description for a webhook.