Amplifybackend

This page documents function available when using the Amplifybackend module, created with @service Amplifybackend.

Index

Main.Amplifybackend.clone_backend

Main.Amplifybackend.create_backend

Main.Amplifybackend.create_backend_api

Main.Amplifybackend.create_backend_auth

Main.Amplifybackend.create_backend_config

Main.Amplifybackend.create_token

Main.Amplifybackend.delete_backend

Main.Amplifybackend.delete_backend_api

Main.Amplifybackend.delete_backend_auth

Main.Amplifybackend.delete_token

Main.Amplifybackend.generate_backend_apimodels

Main.Amplifybackend.get_backend

Main.Amplifybackend.get_backend_api

Main.Amplifybackend.get_backend_apimodels

Main.Amplifybackend.get_backend_auth

Main.Amplifybackend.get_backend_job

Main.Amplifybackend.get_token

Main.Amplifybackend.list_backend_jobs

Main.Amplifybackend.remove_all_backends

Main.Amplifybackend.remove_backend_config

Main.Amplifybackend.update_backend_api

Main.Amplifybackend.update_backend_auth

Main.Amplifybackend.update_backend_config

Main.Amplifybackend.update_backend_job

Documentation

Main.Amplifybackend.clone_backend — Method

clone_backend(app_id, backend_environment_name, target_environment_name)
clone_backend(app_id, backend_environment_name, target_environment_name, params::Dict{String,<:Any})

This operation clones an existing backend.

Arguments

app_id: The app ID.

backend_environment_name: The name of the backend environment.

target_environment_name: The name of the destination backend environment to be created.

Main.Amplifybackend.create_backend — Method

create_backend(app_id, app_name, backend_environment_name)
create_backend(app_id, app_name, backend_environment_name, params::Dict{String,<:Any})

This operation creates a backend for an Amplify app. Backends are automatically created at the time of app creation.

Arguments

app_id: The app ID.

app_name: The name of the app.

backend_environment_name: The name of the backend environment.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"resourceConfig": The resource configuration for the backend creation request.

"resourceName": The name of the resource.

Main.Amplifybackend.create_backend_api — Method

create_backend_api(app_id, backend_environment_name, resource_config, resource_name)
create_backend_api(app_id, backend_environment_name, resource_config, resource_name, params::Dict{String,<:Any})

Creates a new backend API resource.

Arguments

app_id: The app ID.

backend_environment_name: The name of the backend environment.

resource_config: The resource configuration for this request.

resource_name: The name of this resource.

Main.Amplifybackend.create_backend_auth — Method

create_backend_auth(app_id, backend_environment_name, resource_config, resource_name)
create_backend_auth(app_id, backend_environment_name, resource_config, resource_name, params::Dict{String,<:Any})

Creates a new backend authentication resource.

Arguments

app_id: The app ID.

backend_environment_name: The name of the backend environment.

resource_config: The resource configuration for this request object.

resource_name: The name of this resource.

Main.Amplifybackend.create_backend_config — Method

create_backend_config(app_id)
create_backend_config(app_id, params::Dict{String,<:Any})

Creates a config object for a backend.

Arguments

app_id: The app ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"backendManagerAppId": The app ID for the backend manager.

Main.Amplifybackend.create_token — Method

create_token(app_id)
create_token(app_id, params::Dict{String,<:Any})

Generates a one-time challenge code to authenticate a user into your Amplify Admin UI.

Arguments

app_id: The app ID.

Main.Amplifybackend.delete_backend — Method

delete_backend(app_id, backend_environment_name)
delete_backend(app_id, backend_environment_name, params::Dict{String,<:Any})

Removes an existing environment from your Amplify project.

Arguments

app_id: The app ID.

backend_environment_name: The name of the backend environment.

Main.Amplifybackend.delete_backend_api — Method

delete_backend_api(app_id, backend_environment_name, resource_name)
delete_backend_api(app_id, backend_environment_name, resource_name, params::Dict{String,<:Any})

Deletes an existing backend API resource.

Arguments

app_id: The app ID.

backend_environment_name: The name of the backend environment.

resource_name: The name of this resource.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"resourceConfig": Defines the resource configuration for the data model in your Amplify project.

Main.Amplifybackend.delete_backend_auth — Method

delete_backend_auth(app_id, backend_environment_name, resource_name)
delete_backend_auth(app_id, backend_environment_name, resource_name, params::Dict{String,<:Any})

Deletes an existing backend authentication resource.

Arguments

app_id: The app ID.

backend_environment_name: The name of the backend environment.

resource_name: The name of this resource.

Main.Amplifybackend.delete_token — Method

delete_token(app_id, session_id)
delete_token(app_id, session_id, params::Dict{String,<:Any})

Deletes the challenge token based on the given appId and sessionId.

Arguments

app_id: The app ID.

session_id: The session ID.

Main.Amplifybackend.generate_backend_apimodels — Method

generate_backend_apimodels(app_id, backend_environment_name, resource_name)
generate_backend_apimodels(app_id, backend_environment_name, resource_name, params::Dict{String,<:Any})

Generates a model schema for an existing backend API resource.

Arguments

app_id: The app ID.

backend_environment_name: The name of the backend environment.

resource_name: The name of this resource.

Main.Amplifybackend.get_backend — Method

get_backend(app_id)
get_backend(app_id, params::Dict{String,<:Any})

Provides project-level details for your Amplify UI project.

Arguments

app_id: The app ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"backendEnvironmentName": The name of the backend environment.

Main.Amplifybackend.get_backend_api — Method

get_backend_api(app_id, backend_environment_name, resource_name)
get_backend_api(app_id, backend_environment_name, resource_name, params::Dict{String,<:Any})

Gets the details for a backend API.

Arguments

app_id: The app ID.

backend_environment_name: The name of the backend environment.

resource_name: The name of this resource.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"resourceConfig": Defines the resource configuration for the data model in your Amplify project.

Main.Amplifybackend.get_backend_apimodels — Method

get_backend_apimodels(app_id, backend_environment_name, resource_name)
get_backend_apimodels(app_id, backend_environment_name, resource_name, params::Dict{String,<:Any})

Generates a model schema for existing backend API resource.

Arguments

app_id: The app ID.

backend_environment_name: The name of the backend environment.

resource_name: The name of this resource.

Main.Amplifybackend.get_backend_auth — Method

get_backend_auth(app_id, backend_environment_name, resource_name)
get_backend_auth(app_id, backend_environment_name, resource_name, params::Dict{String,<:Any})

Gets backend auth details.

Arguments

app_id: The app ID.

backend_environment_name: The name of the backend environment.

resource_name: The name of this resource.

Main.Amplifybackend.get_backend_job — Method

get_backend_job(app_id, backend_environment_name, job_id)
get_backend_job(app_id, backend_environment_name, job_id, params::Dict{String,<:Any})

Returns information about a specific job.

Arguments

app_id: The app ID.

backend_environment_name: The name of the backend environment.

job_id: The ID for the job.

Main.Amplifybackend.get_token — Method

get_token(app_id, session_id)
get_token(app_id, session_id, params::Dict{String,<:Any})

Gets the challenge token based on the given appId and sessionId.

Arguments

app_id: The app ID.

session_id: The session ID.

Main.Amplifybackend.list_backend_jobs — Method

list_backend_jobs(app_id, backend_environment_name)
list_backend_jobs(app_id, backend_environment_name, params::Dict{String,<:Any})

Lists the jobs for the backend of an Amplify app.

Arguments

app_id: The app ID.

backend_environment_name: The name of the backend environment.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"jobId": The ID for the job.

"maxResults": The maximum number of results that you want in the response.

"nextToken": The token for the next set of results.

"operation": Filters the list of response objects to include only those with the specified operation name.

"status": Filters the list of response objects to include only those with the specified status.

Main.Amplifybackend.remove_all_backends — Method

remove_all_backends(app_id)
remove_all_backends(app_id, params::Dict{String,<:Any})

Removes all backend environments from your Amplify project.

Arguments

app_id: The app ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"cleanAmplifyApp": Cleans up the Amplify Console app if this value is set to true.

Main.Amplifybackend.remove_backend_config — Method

remove_backend_config(app_id)
remove_backend_config(app_id, params::Dict{String,<:Any})

Removes the AWS resources that are required to access the Amplify Admin UI.

Arguments

app_id: The app ID.

Main.Amplifybackend.update_backend_api — Method

update_backend_api(app_id, backend_environment_name, resource_name)
update_backend_api(app_id, backend_environment_name, resource_name, params::Dict{String,<:Any})

Updates an existing backend API resource.

Arguments

app_id: The app ID.

backend_environment_name: The name of the backend environment.

resource_name: The name of this resource.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"resourceConfig": Defines the resource configuration for the data model in your Amplify project.

Main.Amplifybackend.update_backend_auth — Method

update_backend_auth(app_id, backend_environment_name, resource_config, resource_name)
update_backend_auth(app_id, backend_environment_name, resource_config, resource_name, params::Dict{String,<:Any})

Updates an existing backend authentication resource.

Arguments

app_id: The app ID.

backend_environment_name: The name of the backend environment.

resource_config: The resource configuration for this request object.

resource_name: The name of this resource.

Main.Amplifybackend.update_backend_config — Method

update_backend_config(app_id)
update_backend_config(app_id, params::Dict{String,<:Any})

Updates the AWS resources that are required to access the Amplify Admin UI.

Arguments

app_id: The app ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"loginAuthConfig": Describes the Amazon Cognito configuration for Admin UI access.

Main.Amplifybackend.update_backend_job — Method

update_backend_job(app_id, backend_environment_name, job_id)
update_backend_job(app_id, backend_environment_name, job_id, params::Dict{String,<:Any})

Updates a specific job.

Arguments

app_id: The app ID.

backend_environment_name: The name of the backend environment.

job_id: The ID for the job.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"operation": Filters the list of response objects to include only those with the specified operation name.

"status": Filters the list of response objects to include only those with the specified status.