Amplifyuibuilder

This page documents function available when using the Amplifyuibuilder module, created with @service Amplifyuibuilder.

Index

Main.Amplifyuibuilder.create_component

Main.Amplifyuibuilder.create_form

Main.Amplifyuibuilder.create_theme

Main.Amplifyuibuilder.delete_component

Main.Amplifyuibuilder.delete_form

Main.Amplifyuibuilder.delete_theme

Main.Amplifyuibuilder.exchange_code_for_token

Main.Amplifyuibuilder.export_components

Main.Amplifyuibuilder.export_forms

Main.Amplifyuibuilder.export_themes

Main.Amplifyuibuilder.get_codegen_job

Main.Amplifyuibuilder.get_component

Main.Amplifyuibuilder.get_form

Main.Amplifyuibuilder.get_metadata

Main.Amplifyuibuilder.get_theme

Main.Amplifyuibuilder.list_codegen_jobs

Main.Amplifyuibuilder.list_components

Main.Amplifyuibuilder.list_forms

Main.Amplifyuibuilder.list_tags_for_resource

Main.Amplifyuibuilder.list_themes

Main.Amplifyuibuilder.put_metadata_flag

Main.Amplifyuibuilder.refresh_token

Main.Amplifyuibuilder.start_codegen_job

Main.Amplifyuibuilder.tag_resource

Main.Amplifyuibuilder.untag_resource

Main.Amplifyuibuilder.update_component

Main.Amplifyuibuilder.update_form

Main.Amplifyuibuilder.update_theme

Documentation

Main.Amplifyuibuilder.create_component — Method

create_component(app_id, component_to_create, environment_name)
create_component(app_id, component_to_create, environment_name, params::Dict{String,<:Any})

Creates a new component for an Amplify app.

Arguments

app_id: The unique ID of the Amplify app to associate with the component.

component_to_create: Represents the configuration of the component to create.

environment_name: The name of the backend environment that is a part of the Amplify app.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": The unique client token.

Main.Amplifyuibuilder.create_form — Method

create_form(app_id, environment_name, form_to_create)
create_form(app_id, environment_name, form_to_create, params::Dict{String,<:Any})

Creates a new form for an Amplify app.

Arguments

app_id: The unique ID of the Amplify app to associate with the form.

environment_name: The name of the backend environment that is a part of the Amplify app.

form_to_create: Represents the configuration of the form to create.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": The unique client token.

Main.Amplifyuibuilder.create_theme — Method

create_theme(app_id, environment_name, theme_to_create)
create_theme(app_id, environment_name, theme_to_create, params::Dict{String,<:Any})

Creates a theme to apply to the components in an Amplify app.

Arguments

app_id: The unique ID of the Amplify app associated with the theme.

environment_name: The name of the backend environment that is a part of the Amplify app.

theme_to_create: Represents the configuration of the theme to create.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": The unique client token.

Main.Amplifyuibuilder.delete_component — Method

delete_component(app_id, environment_name, id)
delete_component(app_id, environment_name, id, params::Dict{String,<:Any})

Deletes a component from an Amplify app.

Arguments

app_id: The unique ID of the Amplify app associated with the component to delete.

environment_name: The name of the backend environment that is a part of the Amplify app.

id: The unique ID of the component to delete.

Main.Amplifyuibuilder.delete_form — Method

delete_form(app_id, environment_name, id)
delete_form(app_id, environment_name, id, params::Dict{String,<:Any})

Deletes a form from an Amplify app.

Arguments

app_id: The unique ID of the Amplify app associated with the form to delete.

environment_name: The name of the backend environment that is a part of the Amplify app.

id: The unique ID of the form to delete.

Main.Amplifyuibuilder.delete_theme — Method

delete_theme(app_id, environment_name, id)
delete_theme(app_id, environment_name, id, params::Dict{String,<:Any})

Deletes a theme from an Amplify app.

Arguments

app_id: The unique ID of the Amplify app associated with the theme to delete.

environment_name: The name of the backend environment that is a part of the Amplify app.

id: The unique ID of the theme to delete.

Main.Amplifyuibuilder.exchange_code_for_token — Method

exchange_code_for_token(provider, request)
exchange_code_for_token(provider, request, params::Dict{String,<:Any})

This is for internal use. Amplify uses this action to exchange an access code for a token.

Arguments

provider: The third-party provider for the token. The only valid value is figma.

request: Describes the configuration of the request.

Main.Amplifyuibuilder.export_components — Method

export_components(app_id, environment_name)
export_components(app_id, environment_name, params::Dict{String,<:Any})

Exports component configurations to code that is ready to integrate into an Amplify app.

Arguments

app_id: The unique ID of the Amplify app to export components to.

environment_name: The name of the backend environment that is a part of the Amplify app.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nextToken": The token to request the next page of results.

Main.Amplifyuibuilder.export_forms — Method

export_forms(app_id, environment_name)
export_forms(app_id, environment_name, params::Dict{String,<:Any})

Exports form configurations to code that is ready to integrate into an Amplify app.

Arguments

app_id: The unique ID of the Amplify app to export forms to.

environment_name: The name of the backend environment that is a part of the Amplify app.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nextToken": The token to request the next page of results.

Main.Amplifyuibuilder.export_themes — Method

export_themes(app_id, environment_name)
export_themes(app_id, environment_name, params::Dict{String,<:Any})

Exports theme configurations to code that is ready to integrate into an Amplify app.

Arguments

app_id: The unique ID of the Amplify app to export the themes to.

environment_name: The name of the backend environment that is part of the Amplify app.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nextToken": The token to request the next page of results.

Main.Amplifyuibuilder.get_codegen_job — Method

get_codegen_job(app_id, environment_name, id)
get_codegen_job(app_id, environment_name, id, params::Dict{String,<:Any})

Returns an existing code generation job.

Arguments

app_id: The unique ID of the Amplify app associated with the code generation job.

environment_name: The name of the backend environment that is a part of the Amplify app associated with the code generation job.

id: The unique ID of the code generation job.

Main.Amplifyuibuilder.get_component — Method

get_component(app_id, environment_name, id)
get_component(app_id, environment_name, id, params::Dict{String,<:Any})

Returns an existing component for an Amplify app.

Arguments

app_id: The unique ID of the Amplify app.

environment_name: The name of the backend environment that is part of the Amplify app.

id: The unique ID of the component.

Main.Amplifyuibuilder.get_form — Method

get_form(app_id, environment_name, id)
get_form(app_id, environment_name, id, params::Dict{String,<:Any})

Returns an existing form for an Amplify app.

Arguments

app_id: The unique ID of the Amplify app.

environment_name: The name of the backend environment that is part of the Amplify app.

id: The unique ID of the form.

Main.Amplifyuibuilder.get_metadata — Method

get_metadata(app_id, environment_name)
get_metadata(app_id, environment_name, params::Dict{String,<:Any})

Returns existing metadata for an Amplify app.

Arguments

app_id: The unique ID of the Amplify app.

environment_name: The name of the backend environment that is part of the Amplify app.

Main.Amplifyuibuilder.get_theme — Method

get_theme(app_id, environment_name, id)
get_theme(app_id, environment_name, id, params::Dict{String,<:Any})

Returns an existing theme for an Amplify app.

Arguments

app_id: The unique ID of the Amplify app.

environment_name: The name of the backend environment that is part of the Amplify app.

id: The unique ID for the theme.

Main.Amplifyuibuilder.list_codegen_jobs — Method

list_codegen_jobs(app_id, environment_name)
list_codegen_jobs(app_id, environment_name, params::Dict{String,<:Any})

Retrieves a list of code generation jobs for a specified Amplify app and backend environment.

Arguments

app_id: The unique ID for the Amplify app.

environment_name: The name of the backend environment that is a part of the Amplify app.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of jobs to retrieve.

"nextToken": The token to request the next page of results.

Main.Amplifyuibuilder.list_components — Method

list_components(app_id, environment_name)
list_components(app_id, environment_name, params::Dict{String,<:Any})

Retrieves a list of components for a specified Amplify app and backend environment.

Arguments

app_id: The unique ID for the Amplify app.

environment_name: The name of the backend environment that is a part of the Amplify app.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of components to retrieve.

"nextToken": The token to request the next page of results.

Main.Amplifyuibuilder.list_forms — Method

list_forms(app_id, environment_name)
list_forms(app_id, environment_name, params::Dict{String,<:Any})

Retrieves a list of forms for a specified Amplify app and backend environment.

Arguments

app_id: The unique ID for the Amplify app.

environment_name: The name of the backend environment that is a part of the Amplify app.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of forms to retrieve.

"nextToken": The token to request the next page of results.

Main.Amplifyuibuilder.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Returns a list of tags for a specified Amazon Resource Name (ARN).

Arguments

resource_arn: The Amazon Resource Name (ARN) to use to list tags.

Main.Amplifyuibuilder.list_themes — Method

list_themes(app_id, environment_name)
list_themes(app_id, environment_name, params::Dict{String,<:Any})

Retrieves a list of themes for a specified Amplify app and backend environment.

Arguments

app_id: The unique ID for the Amplify app.

environment_name: The name of the backend environment that is a part of the Amplify app.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of theme results to return in the response.

"nextToken": The token to request the next page of results.

Main.Amplifyuibuilder.put_metadata_flag — Method

put_metadata_flag(app_id, body, environment_name, feature_name)
put_metadata_flag(app_id, body, environment_name, feature_name, params::Dict{String,<:Any})

Stores the metadata information about a feature on a form.

Arguments

app_id: The unique ID for the Amplify app.

body: The metadata information to store.

environment_name: The name of the backend environment that is part of the Amplify app.

feature_name: The name of the feature associated with the metadata.

Main.Amplifyuibuilder.refresh_token — Method

refresh_token(provider, refresh_token_body)
refresh_token(provider, refresh_token_body, params::Dict{String,<:Any})

This is for internal use. Amplify uses this action to refresh a previously issued access token that might have expired.

Arguments

provider: The third-party provider for the token. The only valid value is figma.

refresh_token_body: Information about the refresh token request.

Main.Amplifyuibuilder.start_codegen_job — Method

start_codegen_job(app_id, codegen_job_to_create, environment_name)
start_codegen_job(app_id, codegen_job_to_create, environment_name, params::Dict{String,<:Any})

Starts a code generation job for a specified Amplify app and backend environment.

Arguments

app_id: The unique ID for the Amplify app.

codegen_job_to_create: The code generation job resource configuration.

environment_name: The name of the backend environment that is a part of the Amplify app.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": The idempotency token used to ensure that the code generation job request completes only once.

Main.Amplifyuibuilder.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Tags the resource with a tag key and value.

Arguments

resource_arn: The Amazon Resource Name (ARN) to use to tag a resource.

tags: A list of tag key value pairs for a specified Amazon Resource Name (ARN).

Main.Amplifyuibuilder.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Untags a resource with a specified Amazon Resource Name (ARN).

Arguments

resource_arn: The Amazon Resource Name (ARN) to use to untag a resource.

tag_keys: The tag keys to use to untag a resource.

Main.Amplifyuibuilder.update_component — Method

update_component(app_id, environment_name, id, updated_component)
update_component(app_id, environment_name, id, updated_component, params::Dict{String,<:Any})

Updates an existing component.

Arguments

app_id: The unique ID for the Amplify app.

environment_name: The name of the backend environment that is part of the Amplify app.

id: The unique ID for the component.

updated_component: The configuration of the updated component.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": The unique client token.

Main.Amplifyuibuilder.update_form — Method

update_form(app_id, environment_name, id, updated_form)
update_form(app_id, environment_name, id, updated_form, params::Dict{String,<:Any})

Updates an existing form.

Arguments

app_id: The unique ID for the Amplify app.

environment_name: The name of the backend environment that is part of the Amplify app.

id: The unique ID for the form.

updated_form: The request accepts the following data in JSON format.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": The unique client token.

Main.Amplifyuibuilder.update_theme — Method

update_theme(app_id, environment_name, id, updated_theme)
update_theme(app_id, environment_name, id, updated_theme, params::Dict{String,<:Any})

Updates an existing theme.

Arguments

app_id: The unique ID for the Amplify app.

environment_name: The name of the backend environment that is part of the Amplify app.

id: The unique ID for the theme.

updated_theme: The configuration of the updated theme.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": The unique client token.