Apigatewayv2

This page documents function available when using the Apigatewayv2 module, created with @service Apigatewayv2.

Index

Main.Apigatewayv2.create_api

Main.Apigatewayv2.create_api_mapping

Main.Apigatewayv2.create_authorizer

Main.Apigatewayv2.create_deployment

Main.Apigatewayv2.create_domain_name

Main.Apigatewayv2.create_integration

Main.Apigatewayv2.create_integration_response

Main.Apigatewayv2.create_model

Main.Apigatewayv2.create_route

Main.Apigatewayv2.create_route_response

Main.Apigatewayv2.create_stage

Main.Apigatewayv2.create_vpc_link

Main.Apigatewayv2.delete_access_log_settings

Main.Apigatewayv2.delete_api

Main.Apigatewayv2.delete_api_mapping

Main.Apigatewayv2.delete_authorizer

Main.Apigatewayv2.delete_cors_configuration

Main.Apigatewayv2.delete_deployment

Main.Apigatewayv2.delete_domain_name

Main.Apigatewayv2.delete_integration

Main.Apigatewayv2.delete_integration_response

Main.Apigatewayv2.delete_model

Main.Apigatewayv2.delete_route

Main.Apigatewayv2.delete_route_request_parameter

Main.Apigatewayv2.delete_route_response

Main.Apigatewayv2.delete_route_settings

Main.Apigatewayv2.delete_stage

Main.Apigatewayv2.delete_vpc_link

Main.Apigatewayv2.export_api

Main.Apigatewayv2.get_api

Main.Apigatewayv2.get_api_mapping

Main.Apigatewayv2.get_api_mappings

Main.Apigatewayv2.get_apis

Main.Apigatewayv2.get_authorizer

Main.Apigatewayv2.get_authorizers

Main.Apigatewayv2.get_deployment

Main.Apigatewayv2.get_deployments

Main.Apigatewayv2.get_domain_name

Main.Apigatewayv2.get_domain_names

Main.Apigatewayv2.get_integration

Main.Apigatewayv2.get_integration_response

Main.Apigatewayv2.get_integration_responses

Main.Apigatewayv2.get_integrations

Main.Apigatewayv2.get_model

Main.Apigatewayv2.get_model_template

Main.Apigatewayv2.get_models

Main.Apigatewayv2.get_route

Main.Apigatewayv2.get_route_response

Main.Apigatewayv2.get_route_responses

Main.Apigatewayv2.get_routes

Main.Apigatewayv2.get_stage

Main.Apigatewayv2.get_stages

Main.Apigatewayv2.get_tags

Main.Apigatewayv2.get_vpc_link

Main.Apigatewayv2.get_vpc_links

Main.Apigatewayv2.import_api

Main.Apigatewayv2.reimport_api

Main.Apigatewayv2.reset_authorizers_cache

Main.Apigatewayv2.tag_resource

Main.Apigatewayv2.untag_resource

Main.Apigatewayv2.update_api

Main.Apigatewayv2.update_api_mapping

Main.Apigatewayv2.update_authorizer

Main.Apigatewayv2.update_deployment

Main.Apigatewayv2.update_domain_name

Main.Apigatewayv2.update_integration

Main.Apigatewayv2.update_integration_response

Main.Apigatewayv2.update_model

Main.Apigatewayv2.update_route

Main.Apigatewayv2.update_route_response

Main.Apigatewayv2.update_stage

Main.Apigatewayv2.update_vpc_link

Documentation

Main.Apigatewayv2.create_api — Method

CreateApi()

Creates an Api resource.

Required Parameters

name: The name of the API.

protocolType: The API protocol.

Optional Parameters

apiKeySelectionExpression: An API key selection expression. Supported only for WebSocket APIs. See API Key Selection Expressions.

corsConfiguration: A CORS configuration. Supported only for HTTP APIs. See Configuring CORS for more information.

credentialsArn: This property is part of quick create. It specifies the credentials required for the integration, if any. For a Lambda integration, three options are available. To specify an IAM Role for API Gateway to assume, use the role's Amazon Resource Name (ARN). To require that the caller's identity be passed through from the request, specify arn:aws:iam:::user/. To use resource-based permissions on supported AWS services, specify null. Currently, this property is not used for HTTP integrations. Supported only for HTTP APIs.

description: The description of the API.

disableExecuteApiEndpoint: Specifies whether clients can invoke your API by using the default execute-api endpoint. By default, clients can invoke your API with the default https://{api_id}.execute-api.{region}.amazonaws.com endpoint. To require that clients use a custom domain name to invoke your API, disable the default endpoint.

disableSchemaValidation: Avoid validating models when creating a deployment. Supported only for WebSocket APIs.

routeKey: This property is part of quick create. If you don't specify a routeKey, a default route of default is created. The default route acts as a catch-all for any request made to your API, for a particular stage. The default route key can't be modified. You can add routes after creating the API, and you can update the route keys of additional routes. Supported only for HTTP APIs.

routeSelectionExpression: The route selection expression for the API. For HTTP APIs, the routeSelectionExpression must be {request.method} {request.path}. If not provided, this will be the default for HTTP APIs. This property is required for WebSocket APIs.

tags: The collection of tags. Each tag element is associated with a given resource.

target: This property is part of quick create. Quick create produces an API with an integration, a default catch-all route, and a default stage which is configured to automatically deploy changes. For HTTP integrations, specify a fully qualified URL. For Lambda integrations, specify a function ARN. The type of the integration will be HTTPPROXY or AWSPROXY, respectively. Supported only for HTTP APIs.

version: A version identifier for the API.

Main.Apigatewayv2.create_api_mapping — Method

CreateApiMapping()

Creates an API mapping.

Required Parameters

apiId: The API identifier.

domainName: The domain name.

stage: The API stage.

Optional Parameters

apiMappingKey: The API mapping key.

Main.Apigatewayv2.create_authorizer — Method

CreateAuthorizer()

Creates an Authorizer for an API.

Required Parameters

apiId: The API identifier.

authorizerType: The authorizer type. Specify REQUEST for a Lambda function using incoming request parameters. Specify JWT to use JSON Web Tokens (supported only for HTTP APIs).

identitySource: The identity source for which authorization is requested. For a REQUEST authorizer, this is optional. The value is a set of one or more mapping expressions of the specified request parameters. The identity source can be headers, query string parameters, stage variables, and context parameters. For example, if an Auth header and a Name query string parameter are defined as identity sources, this value is route.request.header.Auth, route.request.querystring.Name for WebSocket APIs. For HTTP APIs, use selection expressions prefixed with , for example, request.header.Auth, request.querystring.Name. These parameters are used to perform runtime validation for Lambda-based authorizers by verifying all of the identity-related request parameters are present in the request, not null, and non-empty. Only when this is true does the authorizer invoke the authorizer Lambda function. Otherwise, it returns a 401 Unauthorized response without calling the Lambda function. For HTTP APIs, identity sources are also used as the cache key when caching is enabled. To learn more, see Working with AWS Lambda authorizers for HTTP APIs. For JWT, a single entry that specifies where to extract the JSON Web Token (JWT) from inbound requests. Currently only header-based and query parameter-based selections are supported, for example request.header.Authorization.

name: The name of the authorizer.

Optional Parameters

authorizerCredentialsArn: Specifies the required credentials as an IAM role for API Gateway to invoke the authorizer. To specify an IAM role for API Gateway to assume, use the role's Amazon Resource Name (ARN). To use resource-based permissions on the Lambda function, don't specify this parameter. Supported only for REQUEST authorizers.

authorizerPayloadFormatVersion: Specifies the format of the payload sent to an HTTP API Lambda authorizer. Required for HTTP API Lambda authorizers. Supported values are 1.0 and 2.0. To learn more, see Working with AWS Lambda authorizers for HTTP APIs.

authorizerResultTtlInSeconds: The time to live (TTL) for cached authorizer results, in seconds. If it equals 0, authorization caching is disabled. If it is greater than 0, API Gateway caches authorizer responses. The maximum value is 3600, or 1 hour. Supported only for HTTP API Lambda authorizers.

authorizerUri: The authorizer's Uniform Resource Identifier (URI). For REQUEST authorizers, this must be a well-formed Lambda function URI, for example, arn:aws:apigateway:us-west-2:lambda:path/2015-03-31/functions/arn:aws:lambda:us-west-2:{accountid}:function:{lambdafunctionname}/invocations. In general, the URI has this form: arn:aws:apigateway:{region}:lambda:path/{serviceapi} , where {region} is the same as the region hosting the Lambda function, path indicates that the remaining substring in the URI should be treated as the path to the resource, including the initial /. For Lambda functions, this is usually of the form /2015-03-31/functions/[FunctionARN]/invocations. Supported only for REQUEST authorizers.

enableSimpleResponses: Specifies whether a Lambda authorizer returns a response in a simple format. By default, a Lambda authorizer must return an IAM policy. If enabled, the Lambda authorizer can return a boolean value instead of an IAM policy. Supported only for HTTP APIs. To learn more, see Working with AWS Lambda authorizers for HTTP APIs

identityValidationExpression: This parameter is not used.

jwtConfiguration: Represents the configuration of a JWT authorizer. Required for the JWT authorizer type. Supported only for HTTP APIs.

Main.Apigatewayv2.create_deployment — Method

CreateDeployment()

Creates a Deployment for an API.

Required Parameters

apiId: The API identifier.

Optional Parameters

description: The description for the deployment resource.

stageName: The name of the Stage resource for the Deployment resource to create.

Main.Apigatewayv2.create_domain_name — Method

CreateDomainName()

Creates a domain name.

Required Parameters

domainName: The domain name.

Optional Parameters

domainNameConfigurations: The domain name configurations.

mutualTlsAuthentication: The mutual TLS authentication configuration for a custom domain name.

tags: The collection of tags associated with a domain name.

Main.Apigatewayv2.create_integration — Method

CreateIntegration()

Creates an Integration.

Required Parameters

apiId: The API identifier.

integrationType: The integration type of an integration. One of the following: AWS: for integrating the route or method request with an AWS service action, including the Lambda function-invoking action. With the Lambda function-invoking action, this is referred to as the Lambda custom integration. With any other AWS service action, this is known as AWS integration. Supported only for WebSocket APIs. AWSPROXY: for integrating the route or method request with a Lambda function or other AWS service action. This integration is also referred to as a Lambda proxy integration. HTTP: for integrating the route or method request with an HTTP endpoint. This integration is also referred to as the HTTP custom integration. Supported only for WebSocket APIs. HTTPPROXY: for integrating the route or method request with an HTTP endpoint, with the client request passed through as-is. This is also referred to as HTTP proxy integration. For HTTP API private integrations, use an HTTP_PROXY integration. MOCK: for integrating the route or method request with API Gateway as a "loopback" endpoint without invoking any backend. Supported only for WebSocket APIs.

Optional Parameters

connectionId: The ID of the VPC link for a private integration. Supported only for HTTP APIs.

connectionType: The type of the network connection to the integration endpoint. Specify INTERNET for connections through the public routable internet or VPC_LINK for private connections between API Gateway and resources in a VPC. The default value is INTERNET.

contentHandlingStrategy: Supported only for WebSocket APIs. Specifies how to handle response payload content type conversions. Supported values are CONVERTTOBINARY and CONVERTTOTEXT, with the following behaviors: CONVERTTOBINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob. CONVERTTOTEXT: Converts a response payload from a binary blob to a Base64-encoded string. If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

credentialsArn: Specifies the credentials required for the integration, if any. For AWS integrations, three options are available. To specify an IAM Role for API Gateway to assume, use the role's Amazon Resource Name (ARN). To require that the caller's identity be passed through from the request, specify the string arn:aws:iam:::user/. To use resource-based permissions on supported AWS services, specify null.

description: The description of the integration.

integrationMethod: Specifies the integration's HTTP method type.

integrationSubtype: Supported only for HTTP API AWS_PROXY integrations. Specifies the AWS service action to invoke. To learn more, see Integration subtype reference.

integrationUri: For a Lambda integration, specify the URI of a Lambda function. For an HTTP integration, specify a fully-qualified URL. For an HTTP API private integration, specify the ARN of an Application Load Balancer listener, Network Load Balancer listener, or AWS Cloud Map service. If you specify the ARN of an AWS Cloud Map service, API Gateway uses DiscoverInstances to identify resources. You can use query parameters to target specific resources. To learn more, see DiscoverInstances. For private integrations, all resources must be owned by the same AWS account.

passthroughBehavior: Specifies the pass-through behavior for incoming requests based on the Content-Type header in the request, and the available mapping templates specified as the requestTemplates property on the Integration resource. There are three valid values: WHENNOMATCH, WHENNOTEMPLATES, and NEVER. Supported only for WebSocket APIs. WHENNOMATCH passes the request body for unmapped content types through to the integration backend without transformation. NEVER rejects unmapped content types with an HTTP 415 Unsupported Media Type response. WHENNOTEMPLATES allows pass-through when the integration has no content types mapped to templates. However, if there is at least one content type defined, unmapped content types will be rejected with the same HTTP 415 Unsupported Media Type response.

payloadFormatVersion: Specifies the format of the payload sent to an integration. Required for HTTP APIs.

requestParameters: For WebSocket APIs, a key-value map specifying request parameters that are passed from the method request to the backend. The key is an integration request parameter name and the associated value is a method request parameter value or static value that must be enclosed within single quotes and pre-encoded as required by the backend. The method request parameter value must match the pattern of method.request.{location}.{name} , where {location} is querystring, path, or header; and {name} must be a valid and unique method request parameter name. For HTTP API integrations with a specified integrationSubtype, request parameters are a key-value map specifying parameters that are passed to AWS_PROXY integrations. You can provide static values, or map request data, stage variables, or context variables that are evaluated at runtime. To learn more, see Working with AWS service integrations for HTTP APIs. For HTTP API integrations without a specified integrationSubtype request parameters are a key-value map specifying how to transform HTTP requests before sending them to the backend. The key should follow the pattern &lt;action&gt;:&lt;header|querystring|path&gt;.&lt;location&gt; where action can be append, overwrite or remove. For values, you can provide static values, or map request data, stage variables, or context variables that are evaluated at runtime. To learn more, see Transforming API requests and responses.

requestTemplates: Represents a map of Velocity templates that are applied on the request payload based on the value of the Content-Type header sent by the client. The content type value is the key in this map, and the template (as a String) is the value. Supported only for WebSocket APIs.

responseParameters: Supported only for HTTP APIs. You use response parameters to transform the HTTP response from a backend integration before returning the response to clients. Specify a key-value map from a selection key to response parameters. The selection key must be a valid HTTP status code within the range of 200-599. Response parameters are a key-value map. The key must match pattern &lt;action&gt;:&lt;header&gt;.&lt;location&gt; or overwrite.statuscode. The action can be append, overwrite or remove. The value can be a static value, or map to response data, stage variables, or context variables that are evaluated at runtime. To learn more, see Transforming API requests and responses.

templateSelectionExpression: The template selection expression for the integration.

timeoutInMillis: Custom timeout between 50 and 29,000 milliseconds for WebSocket APIs and between 50 and 30,000 milliseconds for HTTP APIs. The default timeout is 29 seconds for WebSocket APIs and 30 seconds for HTTP APIs.

tlsConfig: The TLS configuration for a private integration. If you specify a TLS configuration, private integration traffic uses the HTTPS protocol. Supported only for HTTP APIs.

Main.Apigatewayv2.create_integration_response — Method

CreateIntegrationResponse()

Creates an IntegrationResponses.

Required Parameters

apiId: The API identifier.

integrationId: The integration ID.

integrationResponseKey: The integration response key.

Optional Parameters

contentHandlingStrategy: Specifies how to handle response payload content type conversions. Supported values are CONVERTTOBINARY and CONVERTTOTEXT, with the following behaviors: CONVERTTOBINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob. CONVERTTOTEXT: Converts a response payload from a binary blob to a Base64-encoded string. If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

responseParameters: A key-value map specifying response parameters that are passed to the method response from the backend. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name}, where {name} is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression}, where {name} is a valid and unique response header name and {JSON-expression} is a valid JSON expression without the prefix.

responseTemplates: The collection of response templates for the integration response as a string-to-string map of key-value pairs. Response templates are represented as a key/value map, with a content-type as the key and a template as the value.

templateSelectionExpression: The template selection expression for the integration response. Supported only for WebSocket APIs.

Main.Apigatewayv2.create_model — Method

CreateModel()

Creates a Model for an API.

Required Parameters

apiId: The API identifier.

name: The name of the model. Must be alphanumeric.

schema: The schema for the model. For application/json models, this should be JSON schema draft 4 model.

Optional Parameters

contentType: The content-type for the model, for example, "application/json".

description: The description of the model.

Main.Apigatewayv2.create_route — Method

CreateRoute()

Creates a Route for an API.

Required Parameters

apiId: The API identifier.

routeKey: The route key for the route.

Optional Parameters

apiKeyRequired: Specifies whether an API key is required for the route. Supported only for WebSocket APIs.

authorizationScopes: The authorization scopes supported by this route.

authorizationType: The authorization type for the route. For WebSocket APIs, valid values are NONE for open access, AWSIAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer For HTTP APIs, valid values are NONE for open access, JWT for using JSON Web Tokens, AWSIAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer.

authorizerId: The identifier of the Authorizer resource to be associated with this route. The authorizer identifier is generated by API Gateway when you created the authorizer.

modelSelectionExpression: The model selection expression for the route. Supported only for WebSocket APIs.

operationName: The operation name for the route.

requestModels: The request models for the route. Supported only for WebSocket APIs.

requestParameters: The request parameters for the route. Supported only for WebSocket APIs.

routeResponseSelectionExpression: The route response selection expression for the route. Supported only for WebSocket APIs.

target: The target for the route.

Main.Apigatewayv2.create_route_response — Method

CreateRouteResponse()

Creates a RouteResponse for a Route.

Required Parameters

apiId: The API identifier.

routeId: The route ID.

routeResponseKey: The route response key.

Optional Parameters

modelSelectionExpression: The model selection expression for the route response. Supported only for WebSocket APIs.

responseModels: The response models for the route response.

responseParameters: The route response parameters.

Main.Apigatewayv2.create_stage — Method

CreateStage()

Creates a Stage for an API.

Required Parameters

apiId: The API identifier.

stageName: The name of the stage.

Optional Parameters

accessLogSettings: Settings for logging access in this stage.

autoDeploy: Specifies whether updates to an API automatically trigger a new deployment. The default value is false.

clientCertificateId: The identifier of a client certificate for a Stage. Supported only for WebSocket APIs.

defaultRouteSettings: The default route settings for the stage.

deploymentId: The deployment identifier of the API stage.

description: The description for the API stage.

routeSettings: Route settings for the stage, by routeKey.

stageVariables: A map that defines the stage variables for a Stage. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&amp;=,]+.

tags: The collection of tags. Each tag element is associated with a given resource.

Main.Apigatewayv2.create_vpc_link — Method

CreateVpcLink()

Creates a VPC link.

Required Parameters

name: The name of the VPC link.

subnetIds: A list of subnet IDs to include in the VPC link.

Optional Parameters

securityGroupIds: A list of security group IDs for the VPC link.

tags: A list of tags.

Main.Apigatewayv2.delete_access_log_settings — Method

DeleteAccessLogSettings()

Deletes the AccessLogSettings for a Stage. To disable access logging for a Stage, delete its AccessLogSettings.

Required Parameters

apiId: The API identifier.

stageName: The stage name. Stage names can only contain alphanumeric characters, hyphens, and underscores. Maximum length is 128 characters.

Main.Apigatewayv2.delete_api — Method

DeleteApi()

Deletes an Api resource.

Required Parameters

apiId: The API identifier.

Main.Apigatewayv2.delete_api_mapping — Method

DeleteApiMapping()

Deletes an API mapping.

Required Parameters

apiMappingId: The API mapping identifier.

domainName: The domain name.

Main.Apigatewayv2.delete_authorizer — Method

DeleteAuthorizer()

Deletes an Authorizer.

Required Parameters

apiId: The API identifier.

authorizerId: The authorizer identifier.

Main.Apigatewayv2.delete_cors_configuration — Method

DeleteCorsConfiguration()

Deletes a CORS configuration.

Required Parameters

apiId: The API identifier.

Main.Apigatewayv2.delete_deployment — Method

DeleteDeployment()

Deletes a Deployment.

Required Parameters

apiId: The API identifier.

deploymentId: The deployment ID.

Main.Apigatewayv2.delete_domain_name — Method

DeleteDomainName()

Deletes a domain name.

Required Parameters

domainName: The domain name.

Main.Apigatewayv2.delete_integration — Method

DeleteIntegration()

Deletes an Integration.

Required Parameters

apiId: The API identifier.

integrationId: The integration ID.

Main.Apigatewayv2.delete_integration_response — Method

DeleteIntegrationResponse()

Deletes an IntegrationResponses.

Required Parameters

apiId: The API identifier.

integrationId: The integration ID.

integrationResponseId: The integration response ID.

Main.Apigatewayv2.delete_model — Method

DeleteModel()

Deletes a Model.

Required Parameters

apiId: The API identifier.

modelId: The model ID.

Main.Apigatewayv2.delete_route — Method

DeleteRoute()

Deletes a Route.

Required Parameters

apiId: The API identifier.

routeId: The route ID.

Main.Apigatewayv2.delete_route_request_parameter — Method

DeleteRouteRequestParameter()

Deletes a route request parameter.

Required Parameters

apiId: The API identifier.

requestParameterKey: The route request parameter key.

routeId: The route ID.

Main.Apigatewayv2.delete_route_response — Method

DeleteRouteResponse()

Deletes a RouteResponse.

Required Parameters

apiId: The API identifier.

routeId: The route ID.

routeResponseId: The route response ID.

Main.Apigatewayv2.delete_route_settings — Method

DeleteRouteSettings()

Deletes the RouteSettings for a stage.

Required Parameters

apiId: The API identifier.

routeKey: The route key.

stageName: The stage name. Stage names can only contain alphanumeric characters, hyphens, and underscores. Maximum length is 128 characters.

Main.Apigatewayv2.delete_stage — Method

DeleteStage()

Deletes a Stage.

Required Parameters

apiId: The API identifier.

stageName: The stage name. Stage names can only contain alphanumeric characters, hyphens, and underscores. Maximum length is 128 characters.

Main.Apigatewayv2.delete_vpc_link — Method

DeleteVpcLink()

Deletes a VPC link.

Required Parameters

vpcLinkId: The ID of the VPC link.

Main.Apigatewayv2.export_api — Method

ExportApi()

Required Parameters

apiId: The API identifier.

outputType: The output type of the exported definition file. Valid values are JSON and YAML.

specification: The version of the API specification to use. OAS30, for OpenAPI 3.0, is the only supported value.

Optional Parameters

exportVersion: The version of the API Gateway export algorithm. API Gateway uses the latest version by default. Currently, the only supported version is 1.0.

includeExtensions: Specifies whether to include API Gateway extensions in the exported API definition. API Gateway extensions are included by default.

stageName: The name of the API stage to export. If you don't specify this property, a representation of the latest API configuration is exported.

Main.Apigatewayv2.get_api — Method

GetApi()

Gets an Api resource.

Required Parameters

apiId: The API identifier.

Main.Apigatewayv2.get_api_mapping — Method

GetApiMapping()

Gets an API mapping.

Required Parameters

apiMappingId: The API mapping identifier.

domainName: The domain name.

Main.Apigatewayv2.get_api_mappings — Method

GetApiMappings()

Gets API mappings.

Required Parameters

domainName: The domain name.

Optional Parameters

maxResults: The maximum number of elements to be returned for this resource.

nextToken: The next page of elements from this collection. Not valid for the last element of the collection.

Main.Apigatewayv2.get_apis — Method

GetApis()

Gets a collection of Api resources.

Optional Parameters

maxResults: The maximum number of elements to be returned for this resource.

nextToken: The next page of elements from this collection. Not valid for the last element of the collection.

Main.Apigatewayv2.get_authorizer — Method

GetAuthorizer()

Gets an Authorizer.

Required Parameters

apiId: The API identifier.

authorizerId: The authorizer identifier.

Main.Apigatewayv2.get_authorizers — Method

GetAuthorizers()

Gets the Authorizers for an API.

Required Parameters

apiId: The API identifier.

Optional Parameters

maxResults: The maximum number of elements to be returned for this resource.

nextToken: The next page of elements from this collection. Not valid for the last element of the collection.

Main.Apigatewayv2.get_deployment — Method

GetDeployment()

Gets a Deployment.

Required Parameters

apiId: The API identifier.

deploymentId: The deployment ID.

Main.Apigatewayv2.get_deployments — Method

GetDeployments()

Gets the Deployments for an API.

Required Parameters

apiId: The API identifier.

Optional Parameters

maxResults: The maximum number of elements to be returned for this resource.

nextToken: The next page of elements from this collection. Not valid for the last element of the collection.

Main.Apigatewayv2.get_domain_name — Method

GetDomainName()

Gets a domain name.

Required Parameters

domainName: The domain name.

Main.Apigatewayv2.get_domain_names — Method

GetDomainNames()

Gets the domain names for an AWS account.

Optional Parameters

maxResults: The maximum number of elements to be returned for this resource.

nextToken: The next page of elements from this collection. Not valid for the last element of the collection.

Main.Apigatewayv2.get_integration — Method

GetIntegration()

Gets an Integration.

Required Parameters

apiId: The API identifier.

integrationId: The integration ID.

Main.Apigatewayv2.get_integration_response — Method

GetIntegrationResponse()

Gets an IntegrationResponses.

Required Parameters

apiId: The API identifier.

integrationId: The integration ID.

integrationResponseId: The integration response ID.

Main.Apigatewayv2.get_integration_responses — Method

GetIntegrationResponses()

Gets the IntegrationResponses for an Integration.

Required Parameters

apiId: The API identifier.

integrationId: The integration ID.

Optional Parameters

maxResults: The maximum number of elements to be returned for this resource.

nextToken: The next page of elements from this collection. Not valid for the last element of the collection.

Main.Apigatewayv2.get_integrations — Method

GetIntegrations()

Gets the Integrations for an API.

Required Parameters

apiId: The API identifier.

Optional Parameters

maxResults: The maximum number of elements to be returned for this resource.

nextToken: The next page of elements from this collection. Not valid for the last element of the collection.

Main.Apigatewayv2.get_model — Method

GetModel()

Gets a Model.

Required Parameters

apiId: The API identifier.

modelId: The model ID.

Main.Apigatewayv2.get_model_template — Method

GetModelTemplate()

Gets a model template.

Required Parameters

apiId: The API identifier.

modelId: The model ID.

Main.Apigatewayv2.get_models — Method

GetModels()

Gets the Models for an API.

Required Parameters

apiId: The API identifier.

Optional Parameters

maxResults: The maximum number of elements to be returned for this resource.

nextToken: The next page of elements from this collection. Not valid for the last element of the collection.

Main.Apigatewayv2.get_route — Method

GetRoute()

Gets a Route.

Required Parameters

apiId: The API identifier.

routeId: The route ID.

Main.Apigatewayv2.get_route_response — Method

GetRouteResponse()

Gets a RouteResponse.

Required Parameters

apiId: The API identifier.

routeId: The route ID.

routeResponseId: The route response ID.

Main.Apigatewayv2.get_route_responses — Method

GetRouteResponses()

Gets the RouteResponses for a Route.

Required Parameters

apiId: The API identifier.

routeId: The route ID.

Optional Parameters

maxResults: The maximum number of elements to be returned for this resource.

nextToken: The next page of elements from this collection. Not valid for the last element of the collection.

Main.Apigatewayv2.get_routes — Method

GetRoutes()

Gets the Routes for an API.

Required Parameters

apiId: The API identifier.

Optional Parameters

maxResults: The maximum number of elements to be returned for this resource.

nextToken: The next page of elements from this collection. Not valid for the last element of the collection.

Main.Apigatewayv2.get_stage — Method

GetStage()

Gets a Stage.

Required Parameters

apiId: The API identifier.

stageName: The stage name. Stage names can only contain alphanumeric characters, hyphens, and underscores. Maximum length is 128 characters.

Main.Apigatewayv2.get_stages — Method

GetStages()

Gets the Stages for an API.

Required Parameters

apiId: The API identifier.

Optional Parameters

maxResults: The maximum number of elements to be returned for this resource.

nextToken: The next page of elements from this collection. Not valid for the last element of the collection.

Main.Apigatewayv2.get_tags — Method

GetTags()

Gets a collection of Tag resources.

Required Parameters

resource-arn: The resource ARN for the tag.

Main.Apigatewayv2.get_vpc_link — Method

GetVpcLink()

Gets a VPC link.

Required Parameters

vpcLinkId: The ID of the VPC link.

Main.Apigatewayv2.get_vpc_links — Method

GetVpcLinks()

Gets a collection of VPC links.

Optional Parameters

maxResults: The maximum number of elements to be returned for this resource.

nextToken: The next page of elements from this collection. Not valid for the last element of the collection.

Main.Apigatewayv2.import_api — Method

ImportApi()

Imports an API.

Required Parameters

body: The OpenAPI definition. Supported only for HTTP APIs.

Optional Parameters

basepath: Specifies how to interpret the base path of the API during import. Valid values are ignore, prepend, and split. The default value is ignore. To learn more, see Set the OpenAPI basePath Property. Supported only for HTTP APIs.

failOnWarnings: Specifies whether to rollback the API creation when a warning is encountered. By default, API creation continues if a warning is encountered.

Main.Apigatewayv2.reimport_api — Method

ReimportApi()

Puts an Api resource.

Required Parameters

apiId: The API identifier.

body: The OpenAPI definition. Supported only for HTTP APIs.

Optional Parameters

basepath: Specifies how to interpret the base path of the API during import. Valid values are ignore, prepend, and split. The default value is ignore. To learn more, see Set the OpenAPI basePath Property. Supported only for HTTP APIs.

failOnWarnings: Specifies whether to rollback the API creation when a warning is encountered. By default, API creation continues if a warning is encountered.

Main.Apigatewayv2.reset_authorizers_cache — Method

ResetAuthorizersCache()

Resets all authorizer cache entries on a stage. Supported only for HTTP APIs.

Required Parameters

apiId: The API identifier.

stageName: The stage name. Stage names can contain only alphanumeric characters, hyphens, and underscores, or be default. Maximum length is 128 characters.

Main.Apigatewayv2.tag_resource — Method

TagResource()

Creates a new Tag resource to represent a tag.

Required Parameters

resource-arn: The resource ARN for the tag.

Optional Parameters

tags: The collection of tags. Each tag element is associated with a given resource.

Main.Apigatewayv2.untag_resource — Method

UntagResource()

Deletes a Tag.

Required Parameters

resource-arn: The resource ARN for the tag.

tagKeys: The Tag keys to delete

Main.Apigatewayv2.update_api — Method

UpdateApi()

Updates an Api resource.

Required Parameters

apiId: The API identifier.

Optional Parameters

apiKeySelectionExpression: An API key selection expression. Supported only for WebSocket APIs. See API Key Selection Expressions.

corsConfiguration: A CORS configuration. Supported only for HTTP APIs.

credentialsArn: This property is part of quick create. It specifies the credentials required for the integration, if any. For a Lambda integration, three options are available. To specify an IAM Role for API Gateway to assume, use the role's Amazon Resource Name (ARN). To require that the caller's identity be passed through from the request, specify arn:aws:iam:::user/. To use resource-based permissions on supported AWS services, don't specify this parameter. Currently, this property is not used for HTTP integrations. If provided, this value replaces the credentials associated with the quick create integration. Supported only for HTTP APIs.

description: The description of the API.

disableExecuteApiEndpoint: Specifies whether clients can invoke your API by using the default execute-api endpoint. By default, clients can invoke your API with the default https://{api_id}.execute-api.{region}.amazonaws.com endpoint. To require that clients use a custom domain name to invoke your API, disable the default endpoint.

disableSchemaValidation: Avoid validating models when creating a deployment. Supported only for WebSocket APIs.

name: The name of the API.

routeKey: This property is part of quick create. If not specified, the route created using quick create is kept. Otherwise, this value replaces the route key of the quick create route. Additional routes may still be added after the API is updated. Supported only for HTTP APIs.

routeSelectionExpression: The route selection expression for the API. For HTTP APIs, the routeSelectionExpression must be {request.method} {request.path}. If not provided, this will be the default for HTTP APIs. This property is required for WebSocket APIs.

target: This property is part of quick create. For HTTP integrations, specify a fully qualified URL. For Lambda integrations, specify a function ARN. The type of the integration will be HTTPPROXY or AWSPROXY, respectively. The value provided updates the integration URI and integration type. You can update a quick-created target, but you can't remove it from an API. Supported only for HTTP APIs.

version: A version identifier for the API.

Main.Apigatewayv2.update_api_mapping — Method

UpdateApiMapping()

The API mapping.

Required Parameters

apiId: The API identifier.

apiMappingId: The API mapping identifier.

domainName: The domain name.

Optional Parameters

apiMappingKey: The API mapping key.

stage: The API stage.

Main.Apigatewayv2.update_authorizer — Method

UpdateAuthorizer()

Updates an Authorizer.

Required Parameters

apiId: The API identifier.

authorizerId: The authorizer identifier.

Optional Parameters

authorizerCredentialsArn: Specifies the required credentials as an IAM role for API Gateway to invoke the authorizer. To specify an IAM role for API Gateway to assume, use the role's Amazon Resource Name (ARN). To use resource-based permissions on the Lambda function, don't specify this parameter.

authorizerPayloadFormatVersion: Specifies the format of the payload sent to an HTTP API Lambda authorizer. Required for HTTP API Lambda authorizers. Supported values are 1.0 and 2.0. To learn more, see Working with AWS Lambda authorizers for HTTP APIs.

authorizerResultTtlInSeconds: The time to live (TTL) for cached authorizer results, in seconds. If it equals 0, authorization caching is disabled. If it is greater than 0, API Gateway caches authorizer responses. The maximum value is 3600, or 1 hour. Supported only for HTTP API Lambda authorizers.

authorizerType: The authorizer type. Specify REQUEST for a Lambda function using incoming request parameters. Specify JWT to use JSON Web Tokens (supported only for HTTP APIs).

authorizerUri: The authorizer's Uniform Resource Identifier (URI). For REQUEST authorizers, this must be a well-formed Lambda function URI, for example, arn:aws:apigateway:us-west-2:lambda:path/2015-03-31/functions/arn:aws:lambda:us-west-2:{accountid}:function:{lambdafunctionname}/invocations. In general, the URI has this form: arn:aws:apigateway:{region}:lambda:path/{serviceapi} , where {region} is the same as the region hosting the Lambda function, path indicates that the remaining substring in the URI should be treated as the path to the resource, including the initial /. For Lambda functions, this is usually of the form /2015-03-31/functions/[FunctionARN]/invocations. Supported only for REQUEST authorizers.

enableSimpleResponses: Specifies whether a Lambda authorizer returns a response in a simple format. By default, a Lambda authorizer must return an IAM policy. If enabled, the Lambda authorizer can return a boolean value instead of an IAM policy. Supported only for HTTP APIs. To learn more, see Working with AWS Lambda authorizers for HTTP APIs

identitySource: The identity source for which authorization is requested. For a REQUEST authorizer, this is optional. The value is a set of one or more mapping expressions of the specified request parameters. The identity source can be headers, query string parameters, stage variables, and context parameters. For example, if an Auth header and a Name query string parameter are defined as identity sources, this value is route.request.header.Auth, route.request.querystring.Name for WebSocket APIs. For HTTP APIs, use selection expressions prefixed with , for example, request.header.Auth, request.querystring.Name. These parameters are used to perform runtime validation for Lambda-based authorizers by verifying all of the identity-related request parameters are present in the request, not null, and non-empty. Only when this is true does the authorizer invoke the authorizer Lambda function. Otherwise, it returns a 401 Unauthorized response without calling the Lambda function. For HTTP APIs, identity sources are also used as the cache key when caching is enabled. To learn more, see Working with AWS Lambda authorizers for HTTP APIs. For JWT, a single entry that specifies where to extract the JSON Web Token (JWT) from inbound requests. Currently only header-based and query parameter-based selections are supported, for example request.header.Authorization.

identityValidationExpression: This parameter is not used.

jwtConfiguration: Represents the configuration of a JWT authorizer. Required for the JWT authorizer type. Supported only for HTTP APIs.

name: The name of the authorizer.

Main.Apigatewayv2.update_deployment — Method

UpdateDeployment()

Updates a Deployment.

Required Parameters

apiId: The API identifier.

deploymentId: The deployment ID.

Optional Parameters

description: The description for the deployment resource.

Main.Apigatewayv2.update_domain_name — Method

UpdateDomainName()

Updates a domain name.

Required Parameters

domainName: The domain name.

Optional Parameters

domainNameConfigurations: The domain name configurations.

mutualTlsAuthentication: The mutual TLS authentication configuration for a custom domain name.

Main.Apigatewayv2.update_integration — Method

UpdateIntegration()

Updates an Integration.

Required Parameters

apiId: The API identifier.

integrationId: The integration ID.

Optional Parameters

connectionId: The ID of the VPC link for a private integration. Supported only for HTTP APIs.

connectionType: The type of the network connection to the integration endpoint. Specify INTERNET for connections through the public routable internet or VPC_LINK for private connections between API Gateway and resources in a VPC. The default value is INTERNET.

contentHandlingStrategy: Supported only for WebSocket APIs. Specifies how to handle response payload content type conversions. Supported values are CONVERTTOBINARY and CONVERTTOTEXT, with the following behaviors: CONVERTTOBINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob. CONVERTTOTEXT: Converts a response payload from a binary blob to a Base64-encoded string. If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

credentialsArn: Specifies the credentials required for the integration, if any. For AWS integrations, three options are available. To specify an IAM Role for API Gateway to assume, use the role's Amazon Resource Name (ARN). To require that the caller's identity be passed through from the request, specify the string arn:aws:iam:::user/. To use resource-based permissions on supported AWS services, specify null.

description: The description of the integration

integrationMethod: Specifies the integration's HTTP method type.

integrationSubtype: Supported only for HTTP API AWS_PROXY integrations. Specifies the AWS service action to invoke. To learn more, see Integration subtype reference.

integrationType: The integration type of an integration. One of the following: AWS: for integrating the route or method request with an AWS service action, including the Lambda function-invoking action. With the Lambda function-invoking action, this is referred to as the Lambda custom integration. With any other AWS service action, this is known as AWS integration. Supported only for WebSocket APIs. AWSPROXY: for integrating the route or method request with a Lambda function or other AWS service action. This integration is also referred to as a Lambda proxy integration. HTTP: for integrating the route or method request with an HTTP endpoint. This integration is also referred to as the HTTP custom integration. Supported only for WebSocket APIs. HTTPPROXY: for integrating the route or method request with an HTTP endpoint, with the client request passed through as-is. This is also referred to as HTTP proxy integration. For HTTP API private integrations, use an HTTP_PROXY integration. MOCK: for integrating the route or method request with API Gateway as a "loopback" endpoint without invoking any backend. Supported only for WebSocket APIs.

integrationUri: For a Lambda integration, specify the URI of a Lambda function. For an HTTP integration, specify a fully-qualified URL. For an HTTP API private integration, specify the ARN of an Application Load Balancer listener, Network Load Balancer listener, or AWS Cloud Map service. If you specify the ARN of an AWS Cloud Map service, API Gateway uses DiscoverInstances to identify resources. You can use query parameters to target specific resources. To learn more, see DiscoverInstances. For private integrations, all resources must be owned by the same AWS account.

passthroughBehavior: Specifies the pass-through behavior for incoming requests based on the Content-Type header in the request, and the available mapping templates specified as the requestTemplates property on the Integration resource. There are three valid values: WHENNOMATCH, WHENNOTEMPLATES, and NEVER. Supported only for WebSocket APIs. WHENNOMATCH passes the request body for unmapped content types through to the integration backend without transformation. NEVER rejects unmapped content types with an HTTP 415 Unsupported Media Type response. WHENNOTEMPLATES allows pass-through when the integration has no content types mapped to templates. However, if there is at least one content type defined, unmapped content types will be rejected with the same HTTP 415 Unsupported Media Type response.

payloadFormatVersion: Specifies the format of the payload sent to an integration. Required for HTTP APIs.

requestParameters: For WebSocket APIs, a key-value map specifying request parameters that are passed from the method request to the backend. The key is an integration request parameter name and the associated value is a method request parameter value or static value that must be enclosed within single quotes and pre-encoded as required by the backend. The method request parameter value must match the pattern of method.request.{location}.{name} , where {location} is querystring, path, or header; and {name} must be a valid and unique method request parameter name. For HTTP API integrations with a specified integrationSubtype, request parameters are a key-value map specifying parameters that are passed to AWS_PROXY integrations. You can provide static values, or map request data, stage variables, or context variables that are evaluated at runtime. To learn more, see Working with AWS service integrations for HTTP APIs. For HTTP API integrations, without a specified integrationSubtype request parameters are a key-value map specifying how to transform HTTP requests before sending them to the backend. The key should follow the pattern &lt;action&gt;:&lt;header|querystring|path&gt;.&lt;location&gt; where action can be append, overwrite or remove. For values, you can provide static values, or map request data, stage variables, or context variables that are evaluated at runtime. To learn more, see Transforming API requests and responses.

requestTemplates: Represents a map of Velocity templates that are applied on the request payload based on the value of the Content-Type header sent by the client. The content type value is the key in this map, and the template (as a String) is the value. Supported only for WebSocket APIs.

responseParameters: Supported only for HTTP APIs. You use response parameters to transform the HTTP response from a backend integration before returning the response to clients. Specify a key-value map from a selection key to response parameters. The selection key must be a valid HTTP status code within the range of 200-599. Response parameters are a key-value map. The key must match pattern &lt;action&gt;:&lt;header&gt;.&lt;location&gt; or overwrite.statuscode. The action can be append, overwrite or remove. The value can be a static value, or map to response data, stage variables, or context variables that are evaluated at runtime. To learn more, see Transforming API requests and responses.

templateSelectionExpression: The template selection expression for the integration.

timeoutInMillis: Custom timeout between 50 and 29,000 milliseconds for WebSocket APIs and between 50 and 30,000 milliseconds for HTTP APIs. The default timeout is 29 seconds for WebSocket APIs and 30 seconds for HTTP APIs.

tlsConfig: The TLS configuration for a private integration. If you specify a TLS configuration, private integration traffic uses the HTTPS protocol. Supported only for HTTP APIs.

Main.Apigatewayv2.update_integration_response — Method

UpdateIntegrationResponse()

Updates an IntegrationResponses.

Required Parameters

apiId: The API identifier.

integrationId: The integration ID.

integrationResponseId: The integration response ID.

Optional Parameters

contentHandlingStrategy: Supported only for WebSocket APIs. Specifies how to handle response payload content type conversions. Supported values are CONVERTTOBINARY and CONVERTTOTEXT, with the following behaviors: CONVERTTOBINARY: Converts a response payload from a Base64-encoded string to the corresponding binary blob. CONVERTTOTEXT: Converts a response payload from a binary blob to a Base64-encoded string. If this property is not defined, the response payload will be passed through from the integration response to the route response or method response without modification.

integrationResponseKey: The integration response key.

responseParameters: A key-value map specifying response parameters that are passed to the method response from the backend. The key is a method response header parameter name and the mapped value is an integration response header value, a static value enclosed within a pair of single quotes, or a JSON expression from the integration response body. The mapping key must match the pattern of method.response.header.{name} , where name is a valid and unique header name. The mapped non-static value must match the pattern of integration.response.header.{name} or integration.response.body.{JSON-expression} , where {name} is a valid and unique response header name and {JSON-expression} is a valid JSON expression without the prefix.

responseTemplates: The collection of response templates for the integration response as a string-to-string map of key-value pairs. Response templates are represented as a key/value map, with a content-type as the key and a template as the value.

templateSelectionExpression: The template selection expression for the integration response. Supported only for WebSocket APIs.

Main.Apigatewayv2.update_model — Method

UpdateModel()

Updates a Model.

Required Parameters

apiId: The API identifier.

modelId: The model ID.

Optional Parameters

contentType: The content-type for the model, for example, "application/json".

description: The description of the model.

name: The name of the model.

schema: The schema for the model. For application/json models, this should be JSON schema draft 4 model.

Main.Apigatewayv2.update_route — Method

UpdateRoute()

Updates a Route.

Required Parameters

apiId: The API identifier.

routeId: The route ID.

Optional Parameters

apiKeyRequired: Specifies whether an API key is required for the route. Supported only for WebSocket APIs.

authorizationScopes: The authorization scopes supported by this route.

authorizationType: The authorization type for the route. For WebSocket APIs, valid values are NONE for open access, AWSIAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer For HTTP APIs, valid values are NONE for open access, JWT for using JSON Web Tokens, AWSIAM for using AWS IAM permissions, and CUSTOM for using a Lambda authorizer.

authorizerId: The identifier of the Authorizer resource to be associated with this route. The authorizer identifier is generated by API Gateway when you created the authorizer.

modelSelectionExpression: The model selection expression for the route. Supported only for WebSocket APIs.

operationName: The operation name for the route.

requestModels: The request models for the route. Supported only for WebSocket APIs.

requestParameters: The request parameters for the route. Supported only for WebSocket APIs.

routeKey: The route key for the route.

routeResponseSelectionExpression: The route response selection expression for the route. Supported only for WebSocket APIs.

target: The target for the route.

Main.Apigatewayv2.update_route_response — Method

UpdateRouteResponse()

Updates a RouteResponse.

Required Parameters

apiId: The API identifier.

routeId: The route ID.

routeResponseId: The route response ID.

Optional Parameters

modelSelectionExpression: The model selection expression for the route response. Supported only for WebSocket APIs.

responseModels: The response models for the route response.

responseParameters: The route response parameters.

routeResponseKey: The route response key.

Main.Apigatewayv2.update_stage — Method

UpdateStage()

Updates a Stage.

Required Parameters

apiId: The API identifier.

stageName: The stage name. Stage names can contain only alphanumeric characters, hyphens, and underscores, or be default. Maximum length is 128 characters.

Optional Parameters

accessLogSettings: Settings for logging access in this stage.

autoDeploy: Specifies whether updates to an API automatically trigger a new deployment. The default value is false.

clientCertificateId: The identifier of a client certificate for a Stage.

defaultRouteSettings: The default route settings for the stage.

deploymentId: The deployment identifier for the API stage. Can't be updated if autoDeploy is enabled.

description: The description for the API stage.

routeSettings: Route settings for the stage.

stageVariables: A map that defines the stage variables for a Stage. Variable names can have alphanumeric and underscore characters, and the values must match [A-Za-z0-9-._~:/?#&amp;=,]+.

Main.Apigatewayv2.update_vpc_link — Method

UpdateVpcLink()

Updates a VPC link.

Required Parameters

vpcLinkId: The ID of the VPC link.

Optional Parameters

name: The name of the VPC link.