Appconfig

This page documents function available when using the Appconfig module, created with @service Appconfig.

Index

Main.Appconfig.create_application

Main.Appconfig.create_configuration_profile

Main.Appconfig.create_deployment_strategy

Main.Appconfig.create_environment

Main.Appconfig.create_extension

Main.Appconfig.create_extension_association

Main.Appconfig.create_hosted_configuration_version

Main.Appconfig.delete_application

Main.Appconfig.delete_configuration_profile

Main.Appconfig.delete_deployment_strategy

Main.Appconfig.delete_environment

Main.Appconfig.delete_extension

Main.Appconfig.delete_extension_association

Main.Appconfig.delete_hosted_configuration_version

Main.Appconfig.get_application

Main.Appconfig.get_configuration

Main.Appconfig.get_configuration_profile

Main.Appconfig.get_deployment

Main.Appconfig.get_deployment_strategy

Main.Appconfig.get_environment

Main.Appconfig.get_extension

Main.Appconfig.get_extension_association

Main.Appconfig.get_hosted_configuration_version

Main.Appconfig.list_applications

Main.Appconfig.list_configuration_profiles

Main.Appconfig.list_deployment_strategies

Main.Appconfig.list_deployments

Main.Appconfig.list_environments

Main.Appconfig.list_extension_associations

Main.Appconfig.list_extensions

Main.Appconfig.list_hosted_configuration_versions

Main.Appconfig.list_tags_for_resource

Main.Appconfig.start_deployment

Main.Appconfig.stop_deployment

Main.Appconfig.tag_resource

Main.Appconfig.untag_resource

Main.Appconfig.update_application

Main.Appconfig.update_configuration_profile

Main.Appconfig.update_deployment_strategy

Main.Appconfig.update_environment

Main.Appconfig.update_extension

Main.Appconfig.update_extension_association

Main.Appconfig.validate_configuration

Documentation

Main.Appconfig.create_application — Method

create_application(name)
create_application(name, params::Dict{String,<:Any})

Creates an application. In AppConfig, an application is simply an organizational construct like a folder. This organizational construct has a relationship with some unit of executable code. For example, you could create an application called MyMobileApp to organize and manage configuration data for a mobile application installed by your users.

Arguments

name: A name for the application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A description of the application.

"Tags": Metadata to assign to the application. Tags help organize and categorize your AppConfig resources. Each tag consists of a key and an optional value, both of which you define.

Main.Appconfig.create_configuration_profile — Method

create_configuration_profile(application_id, location_uri, name)
create_configuration_profile(application_id, location_uri, name, params::Dict{String,<:Any})

Creates a configuration profile, which is information that enables AppConfig to access the configuration source. Valid configuration sources include the following: Configuration data in YAML, JSON, and other formats stored in the AppConfig hosted configuration store Configuration data stored as objects in an Amazon Simple Storage Service (Amazon S3) bucket Pipelines stored in CodePipeline Secrets stored in Secrets Manager Standard and secure string parameters stored in Amazon Web Services Systems Manager Parameter Store Configuration data in SSM documents stored in the Systems Manager document store A configuration profile includes the following information: The URI location of the configuration data. The Identity and Access Management (IAM) role that provides access to the configuration data. A validator for the configuration data. Available validators include either a JSON Schema or an Amazon Web Services Lambda function. For more information, see Create a Configuration and a Configuration Profile in the AppConfig User Guide.

Arguments

application_id: The application ID.

location_uri: A URI to locate the configuration. You can specify the following: For the AppConfig hosted configuration store and for feature flags, specify hosted. For an Amazon Web Services Systems Manager Parameter Store parameter, specify either the parameter name in the format ssm-parameter://&lt;parameter name&gt; or the ARN. For an Secrets Manager secret, specify the URI in the following format: secrets-manager://&lt;secret name&gt;. For an Amazon S3 object, specify the URI in the following format: s3://&lt;bucket&gt;/&lt;objectKey&gt; . Here is an example: s3://my-bucket/my-app/us-east-1/my-config.json For an SSM document, specify either the document name in the format ssm-document://&lt;document name&gt; or the Amazon Resource Name (ARN).

name: A name for the configuration profile.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A description of the configuration profile.

"RetrievalRoleArn": The ARN of an IAM role with permission to access the configuration at the specified LocationUri. A retrieval role ARN is not required for configurations stored in the AppConfig hosted configuration store. It is required for all other sources that store your configuration.

"Tags": Metadata to assign to the configuration profile. Tags help organize and categorize your AppConfig resources. Each tag consists of a key and an optional value, both of which you define.

"Type": The type of configurations contained in the profile. AppConfig supports feature flags and freeform configurations. We recommend you create feature flag configurations to enable or disable new features and freeform configurations to distribute configurations to an application. When calling this API, enter one of the following values for Type: AWS.AppConfig.FeatureFlags AWS.Freeform

"Validators": A list of methods for validating the configuration.

Main.Appconfig.create_deployment_strategy — Method

create_deployment_strategy(deployment_duration_in_minutes, growth_factor, name)
create_deployment_strategy(deployment_duration_in_minutes, growth_factor, name, params::Dict{String,<:Any})

Creates a deployment strategy that defines important criteria for rolling out your configuration to the designated targets. A deployment strategy includes the overall duration required, a percentage of targets to receive the deployment during each interval, an algorithm that defines how percentage grows, and bake time.

Arguments

deployment_duration_in_minutes: Total amount of time for a deployment to last.

growth_factor: The percentage of targets to receive a deployed configuration during each interval.

name: A name for the deployment strategy.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A description of the deployment strategy.

"FinalBakeTimeInMinutes": Specifies the amount of time AppConfig monitors for Amazon CloudWatch alarms after the configuration has been deployed to 100% of its targets, before considering the deployment to be complete. If an alarm is triggered during this time, AppConfig rolls back the deployment. You must configure permissions for AppConfig to roll back based on CloudWatch alarms. For more information, see Configuring permissions for rollback based on Amazon CloudWatch alarms in the AppConfig User Guide.

"GrowthType": The algorithm used to define how percentage grows over time. AppConfig supports the following growth types: Linear: For this type, AppConfig processes the deployment by dividing the total number of targets by the value specified for Step percentage. For example, a linear deployment that uses a Step percentage of 10 deploys the configuration to 10 percent of the hosts. After those deployments are complete, the system deploys the configuration to the next 10 percent. This continues until 100% of the targets have successfully received the configuration. Exponential: For this type, AppConfig processes the deployment exponentially using the following formula: G(2^N). In this formula, G is the growth factor specified by the user and N is the number of steps until the configuration is deployed to all targets. For example, if you specify a growth factor of 2, then the system rolls out the configuration as follows: 2(2^0) 2(2^1) 2(2^2) Expressed numerically, the deployment rolls out as follows: 2% of the targets, 4% of the targets, 8% of the targets, and continues until the configuration has been deployed to all targets.

"ReplicateTo": Save the deployment strategy to a Systems Manager (SSM) document.

"Tags": Metadata to assign to the deployment strategy. Tags help organize and categorize your AppConfig resources. Each tag consists of a key and an optional value, both of which you define.

Main.Appconfig.create_environment — Method

create_environment(application_id, name)
create_environment(application_id, name, params::Dict{String,<:Any})

Creates an environment. For each application, you define one or more environments. An environment is a deployment group of AppConfig targets, such as applications in a Beta or Production environment. You can also define environments for application subcomponents such as the Web, Mobile and Back-end components for your application. You can configure Amazon CloudWatch alarms for each environment. The system monitors alarms during a configuration deployment. If an alarm is triggered, the system rolls back the configuration.

Arguments

application_id: The application ID.

name: A name for the environment.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A description of the environment.

"Monitors": Amazon CloudWatch alarms to monitor during the deployment process.

"Tags": Metadata to assign to the environment. Tags help organize and categorize your AppConfig resources. Each tag consists of a key and an optional value, both of which you define.

Main.Appconfig.create_extension — Method

create_extension(actions, name)
create_extension(actions, name, params::Dict{String,<:Any})

Creates an AppConfig extension. An extension augments your ability to inject logic or behavior at different points during the AppConfig workflow of creating or deploying a configuration. You can create your own extensions or use the Amazon Web Services authored extensions provided by AppConfig. For most use cases, to create your own extension, you must create an Lambda function to perform any computation and processing defined in the extension. For more information about extensions, see Working with AppConfig extensions in the AppConfig User Guide.

Arguments

actions: The actions defined in the extension.

name: A name for the extension. Each extension name in your account must be unique. Extension versions use the same name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": Information about the extension.

"Latest-Version-Number": You can omit this field when you create an extension. When you create a new version, specify the most recent current version number. For example, you create version 3, enter 2 for this field.

"Parameters": The parameters accepted by the extension. You specify parameter values when you associate the extension to an AppConfig resource by using the CreateExtensionAssociation API action. For Lambda extension actions, these parameters are included in the Lambda request object.

"Tags": Adds one or more tags for the specified extension. Tags are metadata that help you categorize resources in different ways, for example, by purpose, owner, or environment. Each tag consists of a key and an optional value, both of which you define.

Main.Appconfig.create_extension_association — Method

create_extension_association(extension_identifier, resource_identifier)
create_extension_association(extension_identifier, resource_identifier, params::Dict{String,<:Any})

When you create an extension or configure an Amazon Web Services authored extension, you associate the extension with an AppConfig application, environment, or configuration profile. For example, you can choose to run the AppConfig deployment events to Amazon SNS Amazon Web Services authored extension and receive notifications on an Amazon SNS topic anytime a configuration deployment is started for a specific application. Defining which extension to associate with an AppConfig resource is called an extension association. An extension association is a specified relationship between an extension and an AppConfig resource, such as an application or a configuration profile. For more information about extensions and associations, see Working with AppConfig extensions in the AppConfig User Guide.

Arguments

extension_identifier: The name, the ID, or the Amazon Resource Name (ARN) of the extension.

resource_identifier: The ARN of an application, configuration profile, or environment.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ExtensionVersionNumber": The version number of the extension. If not specified, AppConfig uses the maximum version of the extension.

"Parameters": The parameter names and values defined in the extensions. Extension parameters marked Required must be entered for this field.

"Tags": Adds one or more tags for the specified extension association. Tags are metadata that help you categorize resources in different ways, for example, by purpose, owner, or environment. Each tag consists of a key and an optional value, both of which you define.

Main.Appconfig.create_hosted_configuration_version — Method

create_hosted_configuration_version(application_id, configuration_profile_id, content, content-_type)
create_hosted_configuration_version(application_id, configuration_profile_id, content, content-_type, params::Dict{String,<:Any})

Creates a new configuration in the AppConfig hosted configuration store.

Arguments

application_id: The application ID.

configuration_profile_id: The configuration profile ID.

content: The content of the configuration or the configuration data.

content-_type: A standard MIME type describing the format of the configuration content. For more information, see Content-Type.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A description of the configuration.

"Latest-Version-Number": An optional locking token used to prevent race conditions from overwriting configuration updates when creating a new version. To ensure your data is not overwritten when creating multiple hosted configuration versions in rapid succession, specify the version number of the latest hosted configuration version.

"VersionLabel": An optional, user-defined label for the AppConfig hosted configuration version. This value must contain at least one non-numeric character. For example, "v2.2.0".

Main.Appconfig.delete_application — Method

delete_application(application_id)
delete_application(application_id, params::Dict{String,<:Any})

Deletes an application. Deleting an application does not delete a configuration from a host.

Arguments

application_id: The ID of the application to delete.

Main.Appconfig.delete_configuration_profile — Method

delete_configuration_profile(application_id, configuration_profile_id)
delete_configuration_profile(application_id, configuration_profile_id, params::Dict{String,<:Any})

Deletes a configuration profile. Deleting a configuration profile does not delete a configuration from a host.

Arguments

application_id: The application ID that includes the configuration profile you want to delete.

configuration_profile_id: The ID of the configuration profile you want to delete.

Main.Appconfig.delete_deployment_strategy — Method

delete_deployment_strategy(deployment_strategy_id)
delete_deployment_strategy(deployment_strategy_id, params::Dict{String,<:Any})

Deletes a deployment strategy. Deleting a deployment strategy does not delete a configuration from a host.

Arguments

deployment_strategy_id: The ID of the deployment strategy you want to delete.

Main.Appconfig.delete_environment — Method

delete_environment(application_id, environment_id)
delete_environment(application_id, environment_id, params::Dict{String,<:Any})

Deletes an environment. Deleting an environment does not delete a configuration from a host.

Arguments

application_id: The application ID that includes the environment that you want to delete.

environment_id: The ID of the environment that you want to delete.

Main.Appconfig.delete_extension — Method

delete_extension(extension_identifier)
delete_extension(extension_identifier, params::Dict{String,<:Any})

Deletes an AppConfig extension. You must delete all associations to an extension before you delete the extension.

Arguments

extension_identifier: The name, ID, or Amazon Resource Name (ARN) of the extension you want to delete.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"version": A specific version of an extension to delete. If omitted, the highest version is deleted.

Main.Appconfig.delete_extension_association — Method

delete_extension_association(extension_association_id)
delete_extension_association(extension_association_id, params::Dict{String,<:Any})

Deletes an extension association. This action doesn't delete extensions defined in the association.

Arguments

extension_association_id: The ID of the extension association to delete.

Main.Appconfig.delete_hosted_configuration_version — Method

delete_hosted_configuration_version(application_id, configuration_profile_id, version_number)
delete_hosted_configuration_version(application_id, configuration_profile_id, version_number, params::Dict{String,<:Any})

Deletes a version of a configuration from the AppConfig hosted configuration store.

Arguments

application_id: The application ID.

configuration_profile_id: The configuration profile ID.

version_number: The versions number to delete.

Main.Appconfig.get_application — Method

get_application(application_id)
get_application(application_id, params::Dict{String,<:Any})

Retrieves information about an application.

Arguments

application_id: The ID of the application you want to get.

Main.Appconfig.get_configuration — Method

get_configuration(application, configuration, environment, client_id)
get_configuration(application, configuration, environment, client_id, params::Dict{String,<:Any})

(Deprecated) Retrieves the latest deployed configuration. Note the following important information. This API action is deprecated. Calls to receive configuration data should use the StartConfigurationSession and GetLatestConfiguration APIs instead. GetConfiguration is a priced call. For more information, see Pricing.

Arguments

application: The application to get. Specify either the application name or the application ID.

configuration: The configuration to get. Specify either the configuration name or the configuration ID.

environment: The environment to get. Specify either the environment name or the environment ID.

client_id: The clientId parameter in the following command is a unique, user-specified ID to identify the client for the configuration. This ID enables AppConfig to deploy the configuration in intervals, as defined in the deployment strategy.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"client_configuration_version": The configuration version returned in the most recent GetConfiguration response. AppConfig uses the value of the ClientConfigurationVersion parameter to identify the configuration version on your clients. If you don’t send ClientConfigurationVersion with each call to GetConfiguration, your clients receive the current configuration. You are charged each time your clients receive a configuration. To avoid excess charges, we recommend you use the StartConfigurationSession and GetLatestConfiguration APIs, which track the client configuration version on your behalf. If you choose to continue using GetConfiguration, we recommend that you include the ClientConfigurationVersion value with every call to GetConfiguration. The value to use for ClientConfigurationVersion comes from the ConfigurationVersion attribute returned by GetConfiguration when there is new or updated data, and should be saved for subsequent calls to GetConfiguration. For more information about working with configurations, see Retrieving the Configuration in the AppConfig User Guide.

Main.Appconfig.get_configuration_profile — Method

get_configuration_profile(application_id, configuration_profile_id)
get_configuration_profile(application_id, configuration_profile_id, params::Dict{String,<:Any})

Retrieves information about a configuration profile.

Arguments

application_id: The ID of the application that includes the configuration profile you want to get.

configuration_profile_id: The ID of the configuration profile that you want to get.

Main.Appconfig.get_deployment — Method

get_deployment(application_id, deployment_number, environment_id)
get_deployment(application_id, deployment_number, environment_id, params::Dict{String,<:Any})

Retrieves information about a configuration deployment.

Arguments

application_id: The ID of the application that includes the deployment you want to get.

deployment_number: The sequence number of the deployment.

environment_id: The ID of the environment that includes the deployment you want to get.

Main.Appconfig.get_deployment_strategy — Method

get_deployment_strategy(deployment_strategy_id)
get_deployment_strategy(deployment_strategy_id, params::Dict{String,<:Any})

Retrieves information about a deployment strategy. A deployment strategy defines important criteria for rolling out your configuration to the designated targets. A deployment strategy includes the overall duration required, a percentage of targets to receive the deployment during each interval, an algorithm that defines how percentage grows, and bake time.

Arguments

deployment_strategy_id: The ID of the deployment strategy to get.

Main.Appconfig.get_environment — Method

get_environment(application_id, environment_id)
get_environment(application_id, environment_id, params::Dict{String,<:Any})

Retrieves information about an environment. An environment is a deployment group of AppConfig applications, such as applications in a Production environment or in an EU_Region environment. Each configuration deployment targets an environment. You can enable one or more Amazon CloudWatch alarms for an environment. If an alarm is triggered during a deployment, AppConfig roles back the configuration.

Arguments

application_id: The ID of the application that includes the environment you want to get.

environment_id: The ID of the environment that you want to get.

Main.Appconfig.get_extension — Method

get_extension(extension_identifier)
get_extension(extension_identifier, params::Dict{String,<:Any})

Returns information about an AppConfig extension.

Arguments

extension_identifier: The name, the ID, or the Amazon Resource Name (ARN) of the extension.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"version_number": The extension version number. If no version number was defined, AppConfig uses the highest version.

Main.Appconfig.get_extension_association — Method

get_extension_association(extension_association_id)
get_extension_association(extension_association_id, params::Dict{String,<:Any})

Returns information about an AppConfig extension association. For more information about extensions and associations, see Working with AppConfig extensions in the AppConfig User Guide.

Arguments

extension_association_id: The extension association ID to get.

Main.Appconfig.get_hosted_configuration_version — Method

get_hosted_configuration_version(application_id, configuration_profile_id, version_number)
get_hosted_configuration_version(application_id, configuration_profile_id, version_number, params::Dict{String,<:Any})

Retrieves information about a specific configuration version.

Arguments

application_id: The application ID.

configuration_profile_id: The configuration profile ID.

version_number: The version.

Main.Appconfig.list_applications — Method

list_applications()
list_applications(params::Dict{String,<:Any})

Lists all applications in your Amazon Web Services account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"max_results": The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

"next_token": A token to start the list. Next token is a pagination token generated by AppConfig to describe what page the previous List call ended on. For the first List request, the nextToken should not be set. On subsequent calls, the nextToken parameter should be set to the previous responses nextToken value. Use this token to get the next set of results.

Main.Appconfig.list_configuration_profiles — Method

list_configuration_profiles(application_id)
list_configuration_profiles(application_id, params::Dict{String,<:Any})

Lists the configuration profiles for an application.

Arguments

application_id: The application ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"max_results": The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

"next_token": A token to start the list. Use this token to get the next set of results.

"type": A filter based on the type of configurations that the configuration profile contains. A configuration can be a feature flag or a freeform configuration.

Main.Appconfig.list_deployment_strategies — Method

list_deployment_strategies()
list_deployment_strategies(params::Dict{String,<:Any})

Lists deployment strategies.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"max_results": The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

"next_token": A token to start the list. Use this token to get the next set of results.

Main.Appconfig.list_deployments — Method

list_deployments(application_id, environment_id)
list_deployments(application_id, environment_id, params::Dict{String,<:Any})

Lists the deployments for an environment in descending deployment number order.

Arguments

application_id: The application ID.

environment_id: The environment ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"max_results": The maximum number of items that may be returned for this call. If there are items that have not yet been returned, the response will include a non-null NextToken that you can provide in a subsequent call to get the next set of results.

"next_token": The token returned by a prior call to this operation indicating the next set of results to be returned. If not specified, the operation will return the first set of results.

Main.Appconfig.list_environments — Method

list_environments(application_id)
list_environments(application_id, params::Dict{String,<:Any})

Lists the environments for an application.

Arguments

application_id: The application ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"max_results": The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

"next_token": A token to start the list. Use this token to get the next set of results.

Main.Appconfig.list_extension_associations — Method

list_extension_associations()
list_extension_associations(params::Dict{String,<:Any})

Lists all AppConfig extension associations in the account. For more information about extensions and associations, see Working with AppConfig extensions in the AppConfig User Guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"extension_identifier": The name, the ID, or the Amazon Resource Name (ARN) of the extension.

"extension_version_number": The version number for the extension defined in the association.

"max_results": The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

"next_token": A token to start the list. Use this token to get the next set of results or pass null to get the first set of results.

"resource_identifier": The ARN of an application, configuration profile, or environment.

Main.Appconfig.list_extensions — Method

list_extensions()
list_extensions(params::Dict{String,<:Any})

Lists all custom and Amazon Web Services authored AppConfig extensions in the account. For more information about extensions, see Working with AppConfig extensions in the AppConfig User Guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"max_results": The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

"name": The extension name.

"next_token": A token to start the list. Use this token to get the next set of results.

Main.Appconfig.list_hosted_configuration_versions — Method

list_hosted_configuration_versions(application_id, configuration_profile_id)
list_hosted_configuration_versions(application_id, configuration_profile_id, params::Dict{String,<:Any})

Lists configurations stored in the AppConfig hosted configuration store by version.

Arguments

application_id: The application ID.

configuration_profile_id: The configuration profile ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"max_results": The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.

"next_token": A token to start the list. Use this token to get the next set of results.

"version_label": An optional filter that can be used to specify the version label of an AppConfig hosted configuration version. This parameter supports filtering by prefix using a wildcard, for example "v2*". If you don't specify an asterisk at the end of the value, only an exact match is returned.

Main.Appconfig.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Retrieves the list of key-value tags assigned to the resource.

Arguments

resource_arn: The resource ARN.

Main.Appconfig.start_deployment — Method

start_deployment(application_id, configuration_profile_id, configuration_version, deployment_strategy_id, environment_id)
start_deployment(application_id, configuration_profile_id, configuration_version, deployment_strategy_id, environment_id, params::Dict{String,<:Any})

Starts a deployment.

Arguments

application_id: The application ID.

configuration_profile_id: The configuration profile ID.

configuration_version: The configuration version to deploy. If deploying an AppConfig hosted configuration version, you can specify either the version number or version label.

deployment_strategy_id: The deployment strategy ID.

environment_id: The environment ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A description of the deployment.

"KmsKeyIdentifier": The KMS key identifier (key ID, key alias, or key ARN). AppConfig uses this ID to encrypt the configuration data using a customer managed key.

"Tags": Metadata to assign to the deployment. Tags help organize and categorize your AppConfig resources. Each tag consists of a key and an optional value, both of which you define.

Main.Appconfig.stop_deployment — Method

stop_deployment(application_id, deployment_number, environment_id)
stop_deployment(application_id, deployment_number, environment_id, params::Dict{String,<:Any})

Stops a deployment. This API action works only on deployments that have a status of DEPLOYING. This action moves the deployment to a status of ROLLED_BACK.

Arguments

application_id: The application ID.

deployment_number: The sequence number of the deployment.

environment_id: The environment ID.

Main.Appconfig.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Assigns metadata to an AppConfig resource. Tags help organize and categorize your AppConfig resources. Each tag consists of a key and an optional value, both of which you define. You can specify a maximum of 50 tags for a resource.

Arguments

resource_arn: The ARN of the resource for which to retrieve tags.

tags: The key-value string map. The valid character set is [a-zA-Z+-=._:/]. The tag key can be up to 128 characters and must not start with aws:. The tag value can be up to 256 characters.

Main.Appconfig.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Deletes a tag key and value from an AppConfig resource.

Arguments

resource_arn: The ARN of the resource for which to remove tags.

tag_keys: The tag keys to delete.

Main.Appconfig.update_application — Method

update_application(application_id)
update_application(application_id, params::Dict{String,<:Any})

Updates an application.

Arguments

application_id: The application ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A description of the application.

"Name": The name of the application.

Main.Appconfig.update_configuration_profile — Method

update_configuration_profile(application_id, configuration_profile_id)
update_configuration_profile(application_id, configuration_profile_id, params::Dict{String,<:Any})

Updates a configuration profile.

Arguments

application_id: The application ID.

configuration_profile_id: The ID of the configuration profile.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A description of the configuration profile.

"Name": The name of the configuration profile.

"RetrievalRoleArn": The ARN of an IAM role with permission to access the configuration at the specified LocationUri.

"Validators": A list of methods for validating the configuration.

Main.Appconfig.update_deployment_strategy — Method

update_deployment_strategy(deployment_strategy_id)
update_deployment_strategy(deployment_strategy_id, params::Dict{String,<:Any})

Updates a deployment strategy.

Arguments

deployment_strategy_id: The deployment strategy ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DeploymentDurationInMinutes": Total amount of time for a deployment to last.

"Description": A description of the deployment strategy.

"FinalBakeTimeInMinutes": The amount of time that AppConfig monitors for alarms before considering the deployment to be complete and no longer eligible for automatic rollback.

"GrowthFactor": The percentage of targets to receive a deployed configuration during each interval.

"GrowthType": The algorithm used to define how percentage grows over time. AppConfig supports the following growth types: Linear: For this type, AppConfig processes the deployment by increments of the growth factor evenly distributed over the deployment time. For example, a linear deployment that uses a growth factor of 20 initially makes the configuration available to 20 percent of the targets. After 1/5th of the deployment time has passed, the system updates the percentage to 40 percent. This continues until 100% of the targets are set to receive the deployed configuration. Exponential: For this type, AppConfig processes the deployment exponentially using the following formula: G(2^N). In this formula, G is the growth factor specified by the user and N is the number of steps until the configuration is deployed to all targets. For example, if you specify a growth factor of 2, then the system rolls out the configuration as follows: 2(2^0) 2(2^1) 2(2^2) Expressed numerically, the deployment rolls out as follows: 2% of the targets, 4% of the targets, 8% of the targets, and continues until the configuration has been deployed to all targets.

Main.Appconfig.update_environment — Method

update_environment(application_id, environment_id)
update_environment(application_id, environment_id, params::Dict{String,<:Any})

Updates an environment.

Arguments

application_id: The application ID.

environment_id: The environment ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A description of the environment.

"Monitors": Amazon CloudWatch alarms to monitor during the deployment process.

"Name": The name of the environment.

Main.Appconfig.update_extension — Method

update_extension(extension_identifier)
update_extension(extension_identifier, params::Dict{String,<:Any})

Updates an AppConfig extension. For more information about extensions, see Working with AppConfig extensions in the AppConfig User Guide.

Arguments

extension_identifier: The name, the ID, or the Amazon Resource Name (ARN) of the extension.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Actions": The actions defined in the extension.

"Description": Information about the extension.

"Parameters": One or more parameters for the actions called by the extension.

"VersionNumber": The extension version number.

Main.Appconfig.update_extension_association — Method

update_extension_association(extension_association_id)
update_extension_association(extension_association_id, params::Dict{String,<:Any})

Updates an association. For more information about extensions and associations, see Working with AppConfig extensions in the AppConfig User Guide.

Arguments

extension_association_id: The system-generated ID for the association.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Parameters": The parameter names and values defined in the extension.

Main.Appconfig.validate_configuration — Method

validate_configuration(application_id, configuration_profile_id, configuration_version)
validate_configuration(application_id, configuration_profile_id, configuration_version, params::Dict{String,<:Any})

Uses the validators in a configuration profile to validate a configuration.

Arguments

application_id: The application ID.

configuration_profile_id: The configuration profile ID.

configuration_version: The version of the configuration to validate.