Arc Region Switch

This page documents function available when using the Arc_Region_Switch module, created with @service Arc_Region_Switch.

Index

Main.Arc_Region_Switch.approve_plan_execution_step

Main.Arc_Region_Switch.cancel_plan_execution

Main.Arc_Region_Switch.create_plan

Main.Arc_Region_Switch.delete_plan

Main.Arc_Region_Switch.get_plan

Main.Arc_Region_Switch.get_plan_evaluation_status

Main.Arc_Region_Switch.get_plan_execution

Main.Arc_Region_Switch.get_plan_in_region

Main.Arc_Region_Switch.list_plan_execution_events

Main.Arc_Region_Switch.list_plan_executions

Main.Arc_Region_Switch.list_plans

Main.Arc_Region_Switch.list_plans_in_region

Main.Arc_Region_Switch.list_route53_health_checks

Main.Arc_Region_Switch.list_route53_health_checks_in_region

Main.Arc_Region_Switch.list_tags_for_resource

Main.Arc_Region_Switch.start_plan_execution

Main.Arc_Region_Switch.tag_resource

Main.Arc_Region_Switch.untag_resource

Main.Arc_Region_Switch.update_plan

Main.Arc_Region_Switch.update_plan_execution

Main.Arc_Region_Switch.update_plan_execution_step

Documentation

Main.Arc_Region_Switch.approve_plan_execution_step — Function

approve_plan_execution_step(approval, execution_id, plan_arn, step_name)
approve_plan_execution_step(approval, execution_id, plan_arn, step_name, params::Dict{String,<:Any})

Approves a step in a plan execution that requires manual approval. When you create a plan, you can include approval steps that require manual intervention before the execution can proceed. This operation allows you to provide that approval.

You must specify the plan ARN, execution ID, step name, and approval status. You can also provide an optional comment explaining the approval decision.

Arguments

approval: The status of approval for a plan execution step.

execution_id: The execution identifier of a plan execution.

plan_arn: The Amazon Resource Name (ARN) of the plan.

step_name: The name of a step in a plan execution.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"comment": A comment that you can enter about a plan execution.

Main.Arc_Region_Switch.cancel_plan_execution — Function

cancel_plan_execution(execution_id, plan_arn)
cancel_plan_execution(execution_id, plan_arn, params::Dict{String,<:Any})

Cancels an in-progress plan execution. This operation stops the execution of the plan and prevents any further steps from being processed.

You must specify the plan ARN and execution ID. You can also provide an optional comment explaining why the execution was canceled.

Arguments

execution_id: The execution identifier of a plan execution.

plan_arn: The Amazon Resource Name (ARN) of the plan.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"comment": A comment that you can enter about canceling a plan execution step.

Main.Arc_Region_Switch.create_plan — Function

create_plan(execution_role, name, recovery_approach, regions, workflows)
create_plan(execution_role, name, recovery_approach, regions, workflows, params::Dict{String,<:Any})

Creates a new Region switch plan. A plan defines the steps required to shift traffic from one Amazon Web Services Region to another.

You must specify a name for the plan, the primary Region, and at least one additional Region. You can also provide a description, execution role, recovery time objective, associated alarms, triggers, and workflows that define the steps to execute during a Region switch.

Arguments

execution_role: An execution role is a way to categorize a Region switch plan.

name: The name of a Region switch plan.

recovery_approach: The recovery approach for a Region switch plan, which can be active/active (activeActive) or active/passive (activePassive).

regions: An array that specifies the Amazon Web Services Regions for a Region switch plan. Specify two Regions.

workflows: An array of workflows included in a Region switch plan.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"associatedAlarms": The alarms associated with a Region switch plan.

"description": The description of a Region switch plan.

"primaryRegion": The primary Amazon Web Services Region for the application. This is the Region where the application normally runs before any Region switch occurs.

"recoveryTimeObjectiveMinutes": Optionally, you can specify an recovery time objective for a Region switch plan, in minutes.

"reportConfiguration":

"tags": The tags to apply to the Region switch plan.

"triggers": The triggers associated with a Region switch plan.

Main.Arc_Region_Switch.delete_plan — Function

delete_plan(arn)
delete_plan(arn, params::Dict{String,<:Any})

Deletes a Region switch plan. You must specify the ARN of the plan to delete.

You cannot delete a plan that has an active execution in progress.

Arguments

arn: The Amazon Resource Name (ARN) of the plan.

Main.Arc_Region_Switch.get_plan — Function

get_plan(arn)
get_plan(arn, params::Dict{String,<:Any})

Retrieves detailed information about a Region switch plan. You must specify the ARN of the plan.

Arguments

arn: The Amazon Resource Name (ARN) of the plan.

Main.Arc_Region_Switch.get_plan_evaluation_status — Function

get_plan_evaluation_status(plan_arn)
get_plan_evaluation_status(plan_arn, params::Dict{String,<:Any})

Retrieves the evaluation status of a Region switch plan. The evaluation status provides information about the last time the plan was evaluated and any warnings or issues detected.

Arguments

plan_arn: The Amazon Resource Name (ARN) of the Region switch plan to retrieve evaluation status for.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The number of objects that you want to return with this call.

"nextToken": Specifies that you want to receive the next page of results. Valid only if you received a nextToken response in the previous request. If you did, it indicates that more output is available. Set this parameter to the value provided by the previous call's nextToken response to request the next page of results.

Main.Arc_Region_Switch.get_plan_execution — Function

get_plan_execution(execution_id, plan_arn)
get_plan_execution(execution_id, plan_arn, params::Dict{String,<:Any})

Retrieves detailed information about a specific plan execution. You must specify the plan ARN and execution ID.

Arguments

execution_id: The execution identifier of a plan execution.

plan_arn: The Amazon Resource Name (ARN) of the plan with the execution to retrieve.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The number of objects that you want to return with this call.

"nextToken": Specifies that you want to receive the next page of results. Valid only if you received a nextToken response in the previous request. If you did, it indicates that more output is available. Set this parameter to the value provided by the previous call's nextToken response to request the next page of results.

Main.Arc_Region_Switch.get_plan_in_region — Function

get_plan_in_region(arn)
get_plan_in_region(arn, params::Dict{String,<:Any})

Retrieves information about a Region switch plan in a specific Amazon Web Services Region. This operation is useful for getting Region-specific information about a plan.

Arguments

arn: The Amazon Resource Name (ARN) of the plan in Region.

Main.Arc_Region_Switch.list_plan_execution_events — Function

list_plan_execution_events(execution_id, plan_arn)
list_plan_execution_events(execution_id, plan_arn, params::Dict{String,<:Any})

Lists the events that occurred during a plan execution. These events provide a detailed timeline of the execution process.

Arguments

execution_id: The execution identifier of a plan execution.

plan_arn: The Amazon Resource Name (ARN) of the plan.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The number of objects that you want to return with this call.

"name": The name of the plan execution event.

"nextToken": Specifies that you want to receive the next page of results. Valid only if you received a nextToken response in the previous request. If you did, it indicates that more output is available. Set this parameter to the value provided by the previous call's nextToken response to request the next page of results.

Main.Arc_Region_Switch.list_plan_executions — Function

list_plan_executions(plan_arn)
list_plan_executions(plan_arn, params::Dict{String,<:Any})

Lists the executions of a Region switch plan. This operation returns information about both current and historical executions.

Arguments

plan_arn: The ARN for the plan.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The number of objects that you want to return with this call.

"nextToken": Specifies that you want to receive the next page of results. Valid only if you received a nextToken response in the previous request. If you did, it indicates that more output is available. Set this parameter to the value provided by the previous call's nextToken response to request the next page of results.

"state": The state of the plan execution. For example, the plan execution might be In Progress.

Main.Arc_Region_Switch.list_plans — Function

list_plans()
list_plans(params::Dict{String,<:Any})

Lists all Region switch plans in your Amazon Web Services account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The number of objects that you want to return with this call.

"nextToken": Specifies that you want to receive the next page of results. Valid only if you received a nextToken response in the previous request. If you did, it indicates that more output is available. Set this parameter to the value provided by the previous call's nextToken response to request the next page of results.

Main.Arc_Region_Switch.list_plans_in_region — Function

list_plans_in_region()
list_plans_in_region(params::Dict{String,<:Any})

Lists all Region switch plans in your Amazon Web Services account that are available in the current Amazon Web Services Region.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The number of objects that you want to return with this call.

"nextToken": Specifies that you want to receive the next page of results. Valid only if you received a nextToken response in the previous request. If you did, it indicates that more output is available. Set this parameter to the value provided by the previous call's nextToken response to request the next page of results.

Main.Arc_Region_Switch.list_route53_health_checks — Function

list_route53_health_checks(arn)
list_route53_health_checks(arn, params::Dict{String,<:Any})

List the Amazon Route 53 health checks.

Arguments

arn: The Amazon Resource Name (ARN) of the Amazon Route 53 health check request.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"hostedZoneId": The hosted zone ID for the health checks.

"maxResults": The number of objects that you want to return with this call.

"nextToken": Specifies that you want to receive the next page of results. Valid only if you received a nextToken response in the previous request. If you did, it indicates that more output is available. Set this parameter to the value provided by the previous call's nextToken response to request the next page of results.

"recordName": The record name for the health checks.

Main.Arc_Region_Switch.list_route53_health_checks_in_region — Function

list_route53_health_checks_in_region(arn)
list_route53_health_checks_in_region(arn, params::Dict{String,<:Any})

List the Amazon Route 53 health checks in a specific Amazon Web Services Region.

Arguments

arn: The Amazon Resource Name (ARN) of the Arc Region Switch Plan.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"hostedZoneId": The hosted zone ID for the health checks.

"maxResults": The number of objects that you want to return with this call.

"nextToken": Specifies that you want to receive the next page of results. Valid only if you received a nextToken response in the previous request. If you did, it indicates that more output is available. Set this parameter to the value provided by the previous call's nextToken response to request the next page of results.

"recordName": The record name for the health checks.

Main.Arc_Region_Switch.list_tags_for_resource — Function

list_tags_for_resource(arn)
list_tags_for_resource(arn, params::Dict{String,<:Any})

Lists the tags attached to a Region switch resource.

Arguments

arn: The Amazon Resource Name (ARN) of the resource.

Main.Arc_Region_Switch.start_plan_execution — Function

start_plan_execution(action, plan_arn, target_region)
start_plan_execution(action, plan_arn, target_region, params::Dict{String,<:Any})

Starts the execution of a Region switch plan. You can execute a plan in either graceful or ungraceful mode.

Specifing ungraceful mode either changes the behavior of the execution blocks in a workflow or skips specific execution blocks.

Arguments

action: The action to perform. Valid values are activate (to shift traffic to the target Region) or deactivate (to shift traffic away from the target Region).

plan_arn: The Amazon Resource Name (ARN) of the plan to execute.

target_region: The Amazon Web Services Region to target with this execution. This is the Region that traffic will be shifted to or from, depending on the action.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"comment": An optional comment explaining why the plan execution is being started.

"latestVersion": A boolean value indicating whether to use the latest version of the plan. If set to false, you must specify a specific version.

"mode": The plan execution mode. Valid values are graceful, for starting the execution in graceful mode, or ungraceful, for starting the execution in ungraceful mode.

"recoveryExecutionId": The execution identifier of the recovery execution that ran in the opposite region post-recovery is ran in. Required when starting a post-recovery execution.

Main.Arc_Region_Switch.tag_resource — Function

tag_resource(arn, tags)
tag_resource(arn, tags, params::Dict{String,<:Any})

Adds or updates tags for a Region switch resource. You can assign metadata to your resources in the form of tags, which are key-value pairs.

Arguments

arn: The Amazon Resource Name (ARN) for a tag that you add to a resource.

tags: Tags that you add to a resource. You can add a maximum of 50 tags in Region switch.

Main.Arc_Region_Switch.untag_resource — Function

untag_resource(arn, resource_tag_keys)
untag_resource(arn, resource_tag_keys, params::Dict{String,<:Any})

Removes tags from a Region switch resource.

Arguments

arn: The Amazon Resource Name (ARN) for a tag you remove a resource from.

resource_tag_keys: Tag keys that you remove from a resource.

Main.Arc_Region_Switch.update_plan — Function

update_plan(arn, execution_role, workflows)
update_plan(arn, execution_role, workflows, params::Dict{String,<:Any})

Updates an existing Region switch plan. You can modify the plan's description, workflows, execution role, recovery time objective, associated alarms, and triggers.

Arguments

arn: The Amazon Resource Name (ARN) of the plan.

execution_role: The updated IAM role ARN that grants Region switch the permissions needed to execute the plan steps.

workflows: The updated workflows for the Region switch plan.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"associatedAlarms": The updated CloudWatch alarms associated with the plan.

"description": The updated description for the Region switch plan.

"recoveryTimeObjectiveMinutes": The updated target recovery time objective (RTO) in minutes for the plan.

"reportConfiguration": The updated report configuration for the plan.

"triggers": The updated conditions that can automatically trigger the execution of the plan.

Main.Arc_Region_Switch.update_plan_execution — Function

update_plan_execution(action, execution_id, plan_arn)
update_plan_execution(action, execution_id, plan_arn, params::Dict{String,<:Any})

Updates an in-progress plan execution. This operation allows you to modify certain aspects of the execution, such as adding a comment or changing the action.

Arguments

action: The action specified for a plan execution, for example, Switch to Graceful or Pause.

execution_id: The execution identifier of a plan execution.

plan_arn: The Amazon Resource Name (ARN) of the plan with the execution to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"comment": An optional comment about the plan execution.

Main.Arc_Region_Switch.update_plan_execution_step — Function

update_plan_execution_step(action_to_take, comment, execution_id, plan_arn, step_name)
update_plan_execution_step(action_to_take, comment, execution_id, plan_arn, step_name, params::Dict{String,<:Any})

Updates a specific step in an in-progress plan execution. This operation allows you to modify the step's comment or action.

Arguments

action_to_take: The updated action to take for the step. This can be used to skip or retry a step.

comment: An optional comment about the plan execution.

execution_id: The unique identifier of the plan execution containing the step to update.

plan_arn: The Amazon Resource Name (ARN) of the plan containing the execution step to update.

step_name: The name of the execution step to update.