Auditmanager

This page documents function available when using the Auditmanager module, created with @service Auditmanager.

Index

Main.Auditmanager.associate_assessment_report_evidence_folder

Main.Auditmanager.batch_associate_assessment_report_evidence

Main.Auditmanager.batch_create_delegation_by_assessment

Main.Auditmanager.batch_delete_delegation_by_assessment

Main.Auditmanager.batch_disassociate_assessment_report_evidence

Main.Auditmanager.batch_import_evidence_to_assessment_control

Main.Auditmanager.create_assessment

Main.Auditmanager.create_assessment_framework

Main.Auditmanager.create_assessment_report

Main.Auditmanager.create_control

Main.Auditmanager.delete_assessment

Main.Auditmanager.delete_assessment_framework

Main.Auditmanager.delete_assessment_report

Main.Auditmanager.delete_control

Main.Auditmanager.deregister_account

Main.Auditmanager.deregister_organization_admin_account

Main.Auditmanager.disassociate_assessment_report_evidence_folder

Main.Auditmanager.get_account_status

Main.Auditmanager.get_assessment

Main.Auditmanager.get_assessment_framework

Main.Auditmanager.get_assessment_report_url

Main.Auditmanager.get_change_logs

Main.Auditmanager.get_control

Main.Auditmanager.get_delegations

Main.Auditmanager.get_evidence

Main.Auditmanager.get_evidence_by_evidence_folder

Main.Auditmanager.get_evidence_folder

Main.Auditmanager.get_evidence_folders_by_assessment

Main.Auditmanager.get_evidence_folders_by_assessment_control

Main.Auditmanager.get_organization_admin_account

Main.Auditmanager.get_services_in_scope

Main.Auditmanager.get_settings

Main.Auditmanager.list_assessment_frameworks

Main.Auditmanager.list_assessment_reports

Main.Auditmanager.list_assessments

Main.Auditmanager.list_controls

Main.Auditmanager.list_keywords_for_data_source

Main.Auditmanager.list_notifications

Main.Auditmanager.list_tags_for_resource

Main.Auditmanager.register_account

Main.Auditmanager.register_organization_admin_account

Main.Auditmanager.tag_resource

Main.Auditmanager.untag_resource

Main.Auditmanager.update_assessment

Main.Auditmanager.update_assessment_control

Main.Auditmanager.update_assessment_control_set_status

Main.Auditmanager.update_assessment_framework

Main.Auditmanager.update_assessment_status

Main.Auditmanager.update_control

Main.Auditmanager.update_settings

Main.Auditmanager.validate_assessment_report_integrity

Documentation

Main.Auditmanager.associate_assessment_report_evidence_folder — Method

associate_assessment_report_evidence_folder(assessment_id, evidence_folder_id)
associate_assessment_report_evidence_folder(assessment_id, evidence_folder_id, params::Dict{String,<:Any})

Associates an evidence folder to the specified assessment report in Audit Manager.

Arguments

assessment_id: The identifier for the specified assessment.

evidence_folder_id: The identifier for the folder in which evidence is stored.

Main.Auditmanager.batch_associate_assessment_report_evidence — Method

batch_associate_assessment_report_evidence(assessment_id, evidence_folder_id, evidence_ids)
batch_associate_assessment_report_evidence(assessment_id, evidence_folder_id, evidence_ids, params::Dict{String,<:Any})

Associates a list of evidence to an assessment report in an Audit Manager assessment.

Arguments

assessment_id: The unique identifier for the specified assessment.

evidence_folder_id: The identifier for the folder in which the evidence is stored.

evidence_ids: The list of evidence identifiers.

Main.Auditmanager.batch_create_delegation_by_assessment — Method

batch_create_delegation_by_assessment(assessment_id, create_delegation_requests)
batch_create_delegation_by_assessment(assessment_id, create_delegation_requests, params::Dict{String,<:Any})

Create a batch of delegations for a specified assessment in Audit Manager.

Arguments

assessment_id: The identifier for the specified assessment.

create_delegation_requests: The API request to batch create delegations in Audit Manager.

Main.Auditmanager.batch_delete_delegation_by_assessment — Method

batch_delete_delegation_by_assessment(assessment_id, delegation_ids)
batch_delete_delegation_by_assessment(assessment_id, delegation_ids, params::Dict{String,<:Any})

Deletes the delegations in the specified Audit Manager assessment.

Arguments

assessment_id: The identifier for the specified assessment.

delegation_ids: The identifiers for the specified delegations.

Main.Auditmanager.batch_disassociate_assessment_report_evidence — Method

batch_disassociate_assessment_report_evidence(assessment_id, evidence_folder_id, evidence_ids)
batch_disassociate_assessment_report_evidence(assessment_id, evidence_folder_id, evidence_ids, params::Dict{String,<:Any})

Disassociates a list of evidence from the specified assessment report in Audit Manager.

Arguments

assessment_id: The identifier for the specified assessment.

evidence_folder_id: The identifier for the folder in which evidence is stored.

evidence_ids: The list of evidence identifiers.

Main.Auditmanager.batch_import_evidence_to_assessment_control — Method

batch_import_evidence_to_assessment_control(assessment_id, control_id, control_set_id, manual_evidence)
batch_import_evidence_to_assessment_control(assessment_id, control_id, control_set_id, manual_evidence, params::Dict{String,<:Any})

Uploads one or more pieces of evidence to the specified control in the assessment in Audit Manager.

Arguments

assessment_id: The identifier for the specified assessment.

control_id: The identifier for the specified control.

control_set_id: The identifier for the specified control set.

manual_evidence: The list of manual evidence objects.

Main.Auditmanager.create_assessment — Method

create_assessment(assessment_reports_destination, framework_id, name, roles, scope)
create_assessment(assessment_reports_destination, framework_id, name, roles, scope, params::Dict{String,<:Any})

Creates an assessment in Audit Manager.

Arguments

assessment_reports_destination: The assessment report storage destination for the specified assessment that is being created.

framework_id: The identifier for the specified framework.

name: The name of the assessment to be created.

roles: The list of roles for the specified assessment.

scope:

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The optional description of the assessment to be created.

"tags": The tags associated with the assessment.

Main.Auditmanager.create_assessment_framework — Method

create_assessment_framework(control_sets, name)
create_assessment_framework(control_sets, name, params::Dict{String,<:Any})

Creates a custom framework in Audit Manager.

Arguments

control_sets: The control sets to be associated with the framework.

name: The name of the new custom framework.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"complianceType": The compliance type that the new custom framework supports, such as CIS or HIPAA.

"description": An optional description for the new custom framework.

"tags": The tags associated with the framework.

Main.Auditmanager.create_assessment_report — Method

create_assessment_report(assessment_id, name)
create_assessment_report(assessment_id, name, params::Dict{String,<:Any})

Creates an assessment report for the specified assessment.

Arguments

assessment_id: The identifier for the specified assessment.

name: The name of the new assessment report.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The description of the assessment report.

Main.Auditmanager.create_control — Method

create_control(control_mapping_sources, name)
create_control(control_mapping_sources, name, params::Dict{String,<:Any})

Creates a new custom control in Audit Manager.

Arguments

control_mapping_sources: The data mapping sources for the specified control.

name: The name of the control.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"actionPlanInstructions": The recommended actions to carry out if the control is not fulfilled.

"actionPlanTitle": The title of the action plan for remediating the control.

"description": The description of the control.

"tags": The tags associated with the control.

"testingInformation": The steps to follow to determine if the control has been satisfied.

Main.Auditmanager.delete_assessment — Method

delete_assessment(assessment_id)
delete_assessment(assessment_id, params::Dict{String,<:Any})

Deletes an assessment in Audit Manager.

Arguments

assessment_id: The identifier for the specified assessment.

Main.Auditmanager.delete_assessment_framework — Method

delete_assessment_framework(framework_id)
delete_assessment_framework(framework_id, params::Dict{String,<:Any})

Deletes a custom framework in Audit Manager.

Arguments

framework_id: The identifier for the specified framework.

Main.Auditmanager.delete_assessment_report — Method

delete_assessment_report(assessment_id, assessment_report_id)
delete_assessment_report(assessment_id, assessment_report_id, params::Dict{String,<:Any})

Deletes an assessment report from an assessment in Audit Manager.

Arguments

assessment_id: The identifier for the specified assessment.

assessment_report_id: The unique identifier for the assessment report.

Main.Auditmanager.delete_control — Method

delete_control(control_id)
delete_control(control_id, params::Dict{String,<:Any})

Deletes a custom control in Audit Manager.

Arguments

control_id: The identifier for the specified control.

Main.Auditmanager.deregister_account — Method

deregister_account()
deregister_account(params::Dict{String,<:Any})

Deregisters an account in Audit Manager.

Main.Auditmanager.deregister_organization_admin_account — Method

deregister_organization_admin_account()
deregister_organization_admin_account(params::Dict{String,<:Any})

Removes the specified member account as a delegated administrator for Audit Manager. When you remove a delegated administrator from your Audit Manager settings, or when you deregister a delegated administrator from Organizations, you continue to have access to the evidence that you previously collected under that account. However, Audit Manager will stop collecting and attaching evidence to that delegated administrator account moving forward.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"adminAccountId": The identifier for the specified administrator account.

Main.Auditmanager.disassociate_assessment_report_evidence_folder — Method

disassociate_assessment_report_evidence_folder(assessment_id, evidence_folder_id)
disassociate_assessment_report_evidence_folder(assessment_id, evidence_folder_id, params::Dict{String,<:Any})

Disassociates an evidence folder from the specified assessment report in Audit Manager.

Arguments

assessment_id: The identifier for the specified assessment.

evidence_folder_id: The identifier for the folder in which evidence is stored.

Main.Auditmanager.get_account_status — Method

get_account_status()
get_account_status(params::Dict{String,<:Any})

Returns the registration status of an account in Audit Manager.

Main.Auditmanager.get_assessment — Method

get_assessment(assessment_id)
get_assessment(assessment_id, params::Dict{String,<:Any})

Returns an assessment from Audit Manager.

Arguments

assessment_id: The identifier for the specified assessment.

Main.Auditmanager.get_assessment_framework — Method

get_assessment_framework(framework_id)
get_assessment_framework(framework_id, params::Dict{String,<:Any})

Returns a framework from Audit Manager.

Arguments

framework_id: The identifier for the specified framework.

Main.Auditmanager.get_assessment_report_url — Method

get_assessment_report_url(assessment_id, assessment_report_id)
get_assessment_report_url(assessment_id, assessment_report_id, params::Dict{String,<:Any})

Returns the URL of a specified assessment report in Audit Manager.

Arguments

assessment_id: The identifier for the specified assessment.

assessment_report_id: The identifier for the assessment report.

Main.Auditmanager.get_change_logs — Method

get_change_logs(assessment_id)
get_change_logs(assessment_id, params::Dict{String,<:Any})

Returns a list of changelogs from Audit Manager.

Arguments

assessment_id: The identifier for the specified assessment.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"controlId": The identifier for the specified control.

"controlSetId": The identifier for the specified control set.

"maxResults": Represents the maximum number of results per page, or per API request call.

"nextToken": The pagination token used to fetch the next set of results.

Main.Auditmanager.get_control — Method

get_control(control_id)
get_control(control_id, params::Dict{String,<:Any})

Returns a control from Audit Manager.

Arguments

control_id: The identifier for the specified control.

Main.Auditmanager.get_delegations — Method

get_delegations()
get_delegations(params::Dict{String,<:Any})

Returns a list of delegations from an audit owner to a delegate.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Represents the maximum number of results per page, or per API request call.

"nextToken": The pagination token used to fetch the next set of results.

Main.Auditmanager.get_evidence — Method

get_evidence(assessment_id, control_set_id, evidence_folder_id, evidence_id)
get_evidence(assessment_id, control_set_id, evidence_folder_id, evidence_id, params::Dict{String,<:Any})

Returns evidence from Audit Manager.

Arguments

assessment_id: The identifier for the specified assessment.

control_set_id: The identifier for the specified control set.

evidence_folder_id: The identifier for the folder in which the evidence is stored.

evidence_id: The identifier for the evidence.

Main.Auditmanager.get_evidence_by_evidence_folder — Method

get_evidence_by_evidence_folder(assessment_id, control_set_id, evidence_folder_id)
get_evidence_by_evidence_folder(assessment_id, control_set_id, evidence_folder_id, params::Dict{String,<:Any})

Returns all evidence from a specified evidence folder in Audit Manager.

Arguments

assessment_id: The identifier for the specified assessment.

control_set_id: The identifier for the control set.

evidence_folder_id: The unique identifier for the folder in which the evidence is stored.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Represents the maximum number of results per page, or per API request call.

"nextToken": The pagination token used to fetch the next set of results.

Main.Auditmanager.get_evidence_folder — Method

get_evidence_folder(assessment_id, control_set_id, evidence_folder_id)
get_evidence_folder(assessment_id, control_set_id, evidence_folder_id, params::Dict{String,<:Any})

Returns an evidence folder from the specified assessment in Audit Manager.

Arguments

assessment_id: The identifier for the specified assessment.

control_set_id: The identifier for the specified control set.

evidence_folder_id: The identifier for the folder in which the evidence is stored.

Main.Auditmanager.get_evidence_folders_by_assessment — Method

get_evidence_folders_by_assessment(assessment_id)
get_evidence_folders_by_assessment(assessment_id, params::Dict{String,<:Any})

Returns the evidence folders from a specified assessment in Audit Manager.

Arguments

assessment_id: The identifier for the specified assessment.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Represents the maximum number of results per page, or per API request call.

"nextToken": The pagination token used to fetch the next set of results.

Main.Auditmanager.get_evidence_folders_by_assessment_control — Method

get_evidence_folders_by_assessment_control(assessment_id, control_id, control_set_id)
get_evidence_folders_by_assessment_control(assessment_id, control_id, control_set_id, params::Dict{String,<:Any})

Returns a list of evidence folders associated with a specified control of an assessment in Audit Manager.

Arguments

assessment_id: The identifier for the specified assessment.

control_id: The identifier for the specified control.

control_set_id: The identifier for the specified control set.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Represents the maximum number of results per page, or per API request call.

"nextToken": The pagination token used to fetch the next set of results.

Main.Auditmanager.get_organization_admin_account — Method

get_organization_admin_account()
get_organization_admin_account(params::Dict{String,<:Any})

Returns the name of the delegated Amazon Web Services administrator account for the organization.

Main.Auditmanager.get_services_in_scope — Method

get_services_in_scope()
get_services_in_scope(params::Dict{String,<:Any})

Returns a list of the in-scope Amazon Web Services services for the specified assessment.

Main.Auditmanager.get_settings — Method

get_settings(attribute)
get_settings(attribute, params::Dict{String,<:Any})

Returns the settings for the specified account.

Arguments

attribute: The list of SettingAttribute enum values.

Main.Auditmanager.list_assessment_frameworks — Method

list_assessment_frameworks(framework_type)
list_assessment_frameworks(framework_type, params::Dict{String,<:Any})

Returns a list of the frameworks available in the Audit Manager framework library.

Arguments

framework_type: The type of framework, such as standard or custom.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Represents the maximum number of results per page, or per API request call.

"nextToken": The pagination token used to fetch the next set of results.

Main.Auditmanager.list_assessment_reports — Method

list_assessment_reports()
list_assessment_reports(params::Dict{String,<:Any})

Returns a list of assessment reports created in Audit Manager.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Represents the maximum number of results per page, or per API request call.

"nextToken": The pagination token used to fetch the next set of results.

Main.Auditmanager.list_assessments — Method

list_assessments()
list_assessments(params::Dict{String,<:Any})

Returns a list of current and past assessments from Audit Manager.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Represents the maximum number of results per page, or per API request call.

"nextToken": The pagination token used to fetch the next set of results.

Main.Auditmanager.list_controls — Method

list_controls(control_type)
list_controls(control_type, params::Dict{String,<:Any})

Returns a list of controls from Audit Manager.

Arguments

control_type: The type of control, such as standard or custom.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Represents the maximum number of results per page, or per API request call.

"nextToken": The pagination token used to fetch the next set of results.

Main.Auditmanager.list_keywords_for_data_source — Method

list_keywords_for_data_source(source)
list_keywords_for_data_source(source, params::Dict{String,<:Any})

Returns a list of keywords that pre-mapped to the specified control data source.

Arguments

source: The control mapping data source to which the keywords apply.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Represents the maximum number of results per page, or per API request call.

"nextToken": The pagination token used to fetch the next set of results.

Main.Auditmanager.list_notifications — Method

list_notifications()
list_notifications(params::Dict{String,<:Any})

Returns a list of all Audit Manager notifications.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Represents the maximum number of results per page, or per API request call.

"nextToken": The pagination token used to fetch the next set of results.

Main.Auditmanager.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Returns a list of tags for the specified resource in Audit Manager.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the specified resource.

Main.Auditmanager.register_account — Method

register_account()
register_account(params::Dict{String,<:Any})

Enables Audit Manager for the specified account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"delegatedAdminAccount": The delegated administrator account for Audit Manager.

"kmsKey": The KMS key details.

Main.Auditmanager.register_organization_admin_account — Method

register_organization_admin_account(admin_account_id)
register_organization_admin_account(admin_account_id, params::Dict{String,<:Any})

Enables an account within the organization as the delegated administrator for Audit Manager.

Arguments

admin_account_id: The identifier for the specified delegated administrator account.

Main.Auditmanager.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Tags the specified resource in Audit Manager.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the specified resource.

tags: The tags to be associated with the resource.

Main.Auditmanager.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes a tag from a resource in Audit Manager.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the specified resource.

tag_keys: The name or key of the tag.

Main.Auditmanager.update_assessment — Method

update_assessment(assessment_id, scope)
update_assessment(assessment_id, scope, params::Dict{String,<:Any})

Edits an Audit Manager assessment.

Arguments

assessment_id: The identifier for the specified assessment.

scope: The scope of the specified assessment.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"assessmentDescription": The description of the specified assessment.

"assessmentName": The name of the specified assessment to be updated.

"assessmentReportsDestination": The assessment report storage destination for the specified assessment that is being updated.

"roles": The list of roles for the specified assessment.

Main.Auditmanager.update_assessment_control — Method

update_assessment_control(assessment_id, control_id, control_set_id)
update_assessment_control(assessment_id, control_id, control_set_id, params::Dict{String,<:Any})

Updates a control within an assessment in Audit Manager.

Arguments

assessment_id: The identifier for the specified assessment.

control_id: The identifier for the specified control.

control_set_id: The identifier for the specified control set.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"commentBody": The comment body text for the specified control.

"controlStatus": The status of the specified control.

Main.Auditmanager.update_assessment_control_set_status — Method

update_assessment_control_set_status(assessment_id, comment, control_set_id, status)
update_assessment_control_set_status(assessment_id, comment, control_set_id, status, params::Dict{String,<:Any})

Updates the status of a control set in an Audit Manager assessment.

Arguments

assessment_id: The identifier for the specified assessment.

comment: The comment related to the status update.

control_set_id: The identifier for the specified control set.

status: The status of the control set that is being updated.

Main.Auditmanager.update_assessment_framework — Method

update_assessment_framework(control_sets, framework_id, name)
update_assessment_framework(control_sets, framework_id, name, params::Dict{String,<:Any})

Updates a custom framework in Audit Manager.

Arguments

control_sets: The control sets associated with the framework.

framework_id: The identifier for the specified framework.

name: The name of the framework to be updated.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"complianceType": The compliance type that the new custom framework supports, such as CIS or HIPAA.

"description": The description of the framework that is to be updated.

Main.Auditmanager.update_assessment_status — Method

update_assessment_status(assessment_id, status)
update_assessment_status(assessment_id, status, params::Dict{String,<:Any})

Updates the status of an assessment in Audit Manager.

Arguments

assessment_id: The identifier for the specified assessment.

status: The current status of the specified assessment.

Main.Auditmanager.update_control — Method

update_control(control_id, control_mapping_sources, name)
update_control(control_id, control_mapping_sources, name, params::Dict{String,<:Any})

Updates a custom control in Audit Manager.

Arguments

control_id: The identifier for the specified control.

control_mapping_sources: The data mapping sources for the specified control.

name: The name of the control to be updated.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"actionPlanInstructions": The recommended actions to carry out if the control is not fulfilled.

"actionPlanTitle": The title of the action plan for remediating the control.

"description": The optional description of the control.

"testingInformation": The steps that to follow to determine if the control has been satisfied.

Main.Auditmanager.update_settings — Method

update_settings()
update_settings(params::Dict{String,<:Any})

Updates Audit Manager settings for the current user account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"defaultAssessmentReportsDestination": The default storage destination for assessment reports.

"defaultProcessOwners": A list of the default audit owners.

"kmsKey": The KMS key details.

"snsTopic": The Amazon Simple Notification Service (Amazon SNS) topic to which Audit Manager sends notifications.

Main.Auditmanager.validate_assessment_report_integrity — Method

validate_assessment_report_integrity(s3_relative_path)
validate_assessment_report_integrity(s3_relative_path, params::Dict{String,<:Any})

Validates the integrity of an assessment report in Audit Manager.

Arguments

s3_relative_path: The relative path of the specified Amazon S3 bucket in which the assessment report is stored.