B2Bi

This page documents function available when using the B2Bi module, created with @service B2Bi.

Index

Main.B2Bi.create_capability

Main.B2Bi.create_partnership

Main.B2Bi.create_profile

Main.B2Bi.create_transformer

Main.B2Bi.delete_capability

Main.B2Bi.delete_partnership

Main.B2Bi.delete_profile

Main.B2Bi.delete_transformer

Main.B2Bi.get_capability

Main.B2Bi.get_partnership

Main.B2Bi.get_profile

Main.B2Bi.get_transformer

Main.B2Bi.get_transformer_job

Main.B2Bi.list_capabilities

Main.B2Bi.list_partnerships

Main.B2Bi.list_profiles

Main.B2Bi.list_tags_for_resource

Main.B2Bi.list_transformers

Main.B2Bi.start_transformer_job

Main.B2Bi.tag_resource

Main.B2Bi.test_mapping

Main.B2Bi.test_parsing

Main.B2Bi.untag_resource

Main.B2Bi.update_capability

Main.B2Bi.update_partnership

Main.B2Bi.update_profile

Main.B2Bi.update_transformer

Documentation

Main.B2Bi.create_capability — Method

create_capability(configuration, name, type)
create_capability(configuration, name, type, params::Dict{String,<:Any})

Instantiates a capability based on the specified parameters. A trading capability contains the information required to transform incoming EDI documents into JSON or XML outputs.

Arguments

configuration: Specifies a structure that contains the details for a capability.

name: Specifies the name of the capability, used to identify it.

type: Specifies the type of the capability. Currently, only edi is supported.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Reserved for future use.

"instructionsDocuments": Specifies one or more locations in Amazon S3, each specifying an EDI document that can be used with this capability. Each item contains the name of the bucket and the key, to identify the document's location.

"tags": Specifies the key-value pairs assigned to ARNs that you can use to group and search for resources by type. You can attach this metadata to resources (capabilities, partnerships, and so on) for any purpose.

Main.B2Bi.create_partnership — Method

create_partnership(capabilities, email, name, profile_id)
create_partnership(capabilities, email, name, profile_id, params::Dict{String,<:Any})

Creates a partnership between a customer and a trading partner, based on the supplied parameters. A partnership represents the connection between you and your trading partner. It ties together a profile and one or more trading capabilities.

Arguments

capabilities: Specifies a list of the capabilities associated with this partnership.

email: Specifies the email address associated with this trading partner.

name: Specifies a descriptive name for the partnership.

profile_id: Specifies the unique, system-generated identifier for the profile connected to this partnership.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Reserved for future use.

"phone": Specifies the phone number associated with the partnership.

"tags": Specifies the key-value pairs assigned to ARNs that you can use to group and search for resources by type. You can attach this metadata to resources (capabilities, partnerships, and so on) for any purpose.

Main.B2Bi.create_profile — Method

create_profile(business_name, logging, name, phone)
create_profile(business_name, logging, name, phone, params::Dict{String,<:Any})

Creates a customer profile. You can have up to five customer profiles, each representing a distinct private network. A profile is the mechanism used to create the concept of a private network.

Arguments

business_name: Specifies the name for the business associated with this profile.

logging: Specifies whether or not logging is enabled for this profile.

name: Specifies the name of the profile.

phone: Specifies the phone number associated with the profile.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Reserved for future use.

"email": Specifies the email address associated with this customer profile.

"tags": Specifies the key-value pairs assigned to ARNs that you can use to group and search for resources by type. You can attach this metadata to resources (capabilities, partnerships, and so on) for any purpose.

Main.B2Bi.create_transformer — Method

create_transformer(edi_type, file_format, mapping_template, name)
create_transformer(edi_type, file_format, mapping_template, name, params::Dict{String,<:Any})

Creates a transformer. A transformer describes how to process the incoming EDI documents and extract the necessary information to the output file.

Arguments

edi_type: Specifies the details for the EDI standard that is being used for the transformer. Currently, only X12 is supported. X12 is a set of standards and corresponding messages that define specific business documents.

file_format: Specifies that the currently supported file formats for EDI transformations are JSON and XML.

mapping_template: Specifies the mapping template for the transformer. This template is used to map the parsed EDI file using JSONata or XSLT.

name: Specifies the name of the transformer, used to identify it.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Reserved for future use.

"sampleDocument": Specifies a sample EDI document that is used by a transformer as a guide for processing the EDI data.

"tags": Specifies the key-value pairs assigned to ARNs that you can use to group and search for resources by type. You can attach this metadata to resources (capabilities, partnerships, and so on) for any purpose.

Main.B2Bi.delete_capability — Method

delete_capability(capability_id)
delete_capability(capability_id, params::Dict{String,<:Any})

Deletes the specified capability. A trading capability contains the information required to transform incoming EDI documents into JSON or XML outputs.

Arguments

capability_id: Specifies a system-assigned unique identifier for the capability.

Main.B2Bi.delete_partnership — Method

delete_partnership(partnership_id)
delete_partnership(partnership_id, params::Dict{String,<:Any})

Deletes the specified partnership. A partnership represents the connection between you and your trading partner. It ties together a profile and one or more trading capabilities.

Arguments

partnership_id: Specifies the unique, system-generated identifier for a partnership.

Main.B2Bi.delete_profile — Method

delete_profile(profile_id)
delete_profile(profile_id, params::Dict{String,<:Any})

Deletes the specified profile. A profile is the mechanism used to create the concept of a private network.

Arguments

profile_id: Specifies the unique, system-generated identifier for the profile.

Main.B2Bi.delete_transformer — Method

delete_transformer(transformer_id)
delete_transformer(transformer_id, params::Dict{String,<:Any})

Deletes the specified transformer. A transformer describes how to process the incoming EDI documents and extract the necessary information to the output file.

Arguments

transformer_id: Specifies the system-assigned unique identifier for the transformer.

Main.B2Bi.get_capability — Method

get_capability(capability_id)
get_capability(capability_id, params::Dict{String,<:Any})

Retrieves the details for the specified capability. A trading capability contains the information required to transform incoming EDI documents into JSON or XML outputs.

Arguments

capability_id: Specifies a system-assigned unique identifier for the capability.

Main.B2Bi.get_partnership — Method

get_partnership(partnership_id)
get_partnership(partnership_id, params::Dict{String,<:Any})

Retrieves the details for a partnership, based on the partner and profile IDs specified. A partnership represents the connection between you and your trading partner. It ties together a profile and one or more trading capabilities.

Arguments

partnership_id: Specifies the unique, system-generated identifier for a partnership.

Main.B2Bi.get_profile — Method

get_profile(profile_id)
get_profile(profile_id, params::Dict{String,<:Any})

Retrieves the details for the profile specified by the profile ID. A profile is the mechanism used to create the concept of a private network.

Arguments

profile_id: Specifies the unique, system-generated identifier for the profile.

Main.B2Bi.get_transformer — Method

get_transformer(transformer_id)
get_transformer(transformer_id, params::Dict{String,<:Any})

Retrieves the details for the transformer specified by the transformer ID. A transformer describes how to process the incoming EDI documents and extract the necessary information to the output file.

Arguments

transformer_id: Specifies the system-assigned unique identifier for the transformer.

Main.B2Bi.get_transformer_job — Method

get_transformer_job(transformer_id, transformer_job_id)
get_transformer_job(transformer_id, transformer_job_id, params::Dict{String,<:Any})

Returns the details of the transformer run, based on the Transformer job ID.

Arguments

transformer_id: Specifies the system-assigned unique identifier for the transformer.

transformer_job_id: Specifies the unique, system-generated identifier for a transformer run.

Main.B2Bi.list_capabilities — Method

list_capabilities()
list_capabilities(params::Dict{String,<:Any})

Lists the capabilities associated with your Amazon Web Services account for your current or specified region. A trading capability contains the information required to transform incoming EDI documents into JSON or XML outputs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Specifies the maximum number of capabilities to return.

"nextToken": When additional results are obtained from the command, a NextToken parameter is returned in the output. You can then pass the NextToken parameter in a subsequent command to continue listing additional resources.

Main.B2Bi.list_partnerships — Method

list_partnerships()
list_partnerships(params::Dict{String,<:Any})

Lists the partnerships associated with your Amazon Web Services account for your current or specified region. A partnership represents the connection between you and your trading partner. It ties together a profile and one or more trading capabilities.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Specifies the maximum number of capabilities to return.

"nextToken": When additional results are obtained from the command, a NextToken parameter is returned in the output. You can then pass the NextToken parameter in a subsequent command to continue listing additional resources.

"profileId": Specifies the unique, system-generated identifier for the profile connected to this partnership.

Main.B2Bi.list_profiles — Method

list_profiles()
list_profiles(params::Dict{String,<:Any})

Lists the profiles associated with your Amazon Web Services account for your current or specified region. A profile is the mechanism used to create the concept of a private network.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Specifies the maximum number of profiles to return.

"nextToken": When additional results are obtained from the command, a NextToken parameter is returned in the output. You can then pass the NextToken parameter in a subsequent command to continue listing additional resources.

Main.B2Bi.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists all of the tags associated with the Amazon Resource Name (ARN) that you specify. The resource can be a capability, partnership, profile, or transformer.

Arguments

resource_arn: Requests the tags associated with a particular Amazon Resource Name (ARN). An ARN is an identifier for a specific Amazon Web Services resource, such as a capability, partnership, profile, or transformer.

Main.B2Bi.list_transformers — Method

list_transformers()
list_transformers(params::Dict{String,<:Any})

Lists the available transformers. A transformer describes how to process the incoming EDI documents and extract the necessary information to the output file.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Specifies the number of items to return for the API response.

"nextToken": When additional results are obtained from the command, a NextToken parameter is returned in the output. You can then pass the NextToken parameter in a subsequent command to continue listing additional resources.

Main.B2Bi.start_transformer_job — Method

start_transformer_job(input_file, output_location, transformer_id)
start_transformer_job(input_file, output_location, transformer_id, params::Dict{String,<:Any})

Runs a job, using a transformer, to parse input EDI (electronic data interchange) file into the output structures used by Amazon Web Services B2BI Data Interchange. If you only want to transform EDI (electronic data interchange) documents, you don't need to create profiles, partnerships or capabilities. Just create and configure a transformer, and then run the StartTransformerJob API to process your files.

Arguments

input_file: Specifies the location of the input file for the transformation. The location consists of an Amazon S3 bucket and prefix.

output_location: Specifies the location of the output file for the transformation. The location consists of an Amazon S3 bucket and prefix.

transformer_id: Specifies the system-assigned unique identifier for the transformer.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Reserved for future use.

Main.B2Bi.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Attaches a key-value pair to a resource, as identified by its Amazon Resource Name (ARN). Resources are capability, partnership, profile, transformers and other entities. There is no response returned from this call.

Arguments

resource_arn: Specifies an Amazon Resource Name (ARN) for a specific Amazon Web Services resource, such as a capability, partnership, profile, or transformer.

tags: Specifies the key-value pairs assigned to ARNs that you can use to group and search for resources by type. You can attach this metadata to resources (capabilities, partnerships, and so on) for any purpose.

Main.B2Bi.test_mapping — Method

test_mapping(file_format, input_file_content, mapping_template)
test_mapping(file_format, input_file_content, mapping_template, params::Dict{String,<:Any})

Maps the input file according to the provided template file. The API call downloads the file contents from the Amazon S3 location, and passes the contents in as a string, to the inputFileContent parameter.

Arguments

file_format: Specifies that the currently supported file formats for EDI transformations are JSON and XML.

input_file_content: Specify the contents of the EDI (electronic data interchange) XML or JSON file that is used as input for the transform.

mapping_template: Specifies the mapping template for the transformer. This template is used to map the parsed EDI file using JSONata or XSLT.

Main.B2Bi.test_parsing — Method

test_parsing(edi_type, file_format, input_file)
test_parsing(edi_type, file_format, input_file, params::Dict{String,<:Any})

Parses the input EDI (electronic data interchange) file. The input file has a file size limit of 250 KB.

Arguments

edi_type: Specifies the details for the EDI standard that is being used for the transformer. Currently, only X12 is supported. X12 is a set of standards and corresponding messages that define specific business documents.

file_format: Specifies that the currently supported file formats for EDI transformations are JSON and XML.

input_file: Specifies an S3Location object, which contains the Amazon S3 bucket and prefix for the location of the input file.

Main.B2Bi.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Detaches a key-value pair from the specified resource, as identified by its Amazon Resource Name (ARN). Resources are capability, partnership, profile, transformers and other entities.

Arguments

resource_arn: Specifies an Amazon Resource Name (ARN) for a specific Amazon Web Services resource, such as a capability, partnership, profile, or transformer.

tag_keys: Specifies the key-value pairs assigned to ARNs that you can use to group and search for resources by type. You can attach this metadata to resources (capabilities, partnerships, and so on) for any purpose.

Main.B2Bi.update_capability — Method

update_capability(capability_id)
update_capability(capability_id, params::Dict{String,<:Any})

Updates some of the parameters for a capability, based on the specified parameters. A trading capability contains the information required to transform incoming EDI documents into JSON or XML outputs.

Arguments

capability_id: Specifies a system-assigned unique identifier for the capability.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"configuration": Specifies a structure that contains the details for a capability.

"instructionsDocuments": Specifies one or more locations in Amazon S3, each specifying an EDI document that can be used with this capability. Each item contains the name of the bucket and the key, to identify the document's location.

"name": Specifies a new name for the capability, to replace the existing name.

Main.B2Bi.update_partnership — Method

update_partnership(partnership_id)
update_partnership(partnership_id, params::Dict{String,<:Any})

Updates some of the parameters for a partnership between a customer and trading partner. A partnership represents the connection between you and your trading partner. It ties together a profile and one or more trading capabilities.

Arguments

partnership_id: Specifies the unique, system-generated identifier for a partnership.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"capabilities": List of the capabilities associated with this partnership.

"name": The name of the partnership, used to identify it.

Main.B2Bi.update_profile — Method

update_profile(profile_id)
update_profile(profile_id, params::Dict{String,<:Any})

Updates the specified parameters for a profile. A profile is the mechanism used to create the concept of a private network.

Arguments

profile_id: Specifies the unique, system-generated identifier for the profile.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"businessName": Specifies the name for the business associated with this profile.

"email": Specifies the email address associated with this customer profile.

"name": The name of the profile, used to identify it.

"phone": Specifies the phone number associated with the profile.

Main.B2Bi.update_transformer — Method

update_transformer(transformer_id)
update_transformer(transformer_id, params::Dict{String,<:Any})

Updates the specified parameters for a transformer. A transformer describes how to process the incoming EDI documents and extract the necessary information to the output file.

Arguments

transformer_id: Specifies the system-assigned unique identifier for the transformer.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ediType": Specifies the details for the EDI standard that is being used for the transformer. Currently, only X12 is supported. X12 is a set of standards and corresponding messages that define specific business documents.

"fileFormat": Specifies that the currently supported file formats for EDI transformations are JSON and XML.

"mappingTemplate": Specifies the mapping template for the transformer. This template is used to map the parsed EDI file using JSONata or XSLT.

"name": Specify a new name for the transformer, if you want to update it.

"sampleDocument": Specifies a sample EDI document that is used by a transformer as a guide for processing the EDI data.

"status": Specifies the transformer's status. You can update the state of the transformer, from active to inactive, or inactive to active.