Backup

This page documents function available when using the Backup module, created with @service Backup.

Index

Main.Backup.create_backup_plan

Main.Backup.create_backup_selection

Main.Backup.create_backup_vault

Main.Backup.create_framework

Main.Backup.create_report_plan

Main.Backup.delete_backup_plan

Main.Backup.delete_backup_selection

Main.Backup.delete_backup_vault

Main.Backup.delete_backup_vault_access_policy

Main.Backup.delete_backup_vault_notifications

Main.Backup.delete_framework

Main.Backup.delete_recovery_point

Main.Backup.delete_report_plan

Main.Backup.describe_backup_job

Main.Backup.describe_backup_vault

Main.Backup.describe_copy_job

Main.Backup.describe_framework

Main.Backup.describe_global_settings

Main.Backup.describe_protected_resource

Main.Backup.describe_recovery_point

Main.Backup.describe_region_settings

Main.Backup.describe_report_job

Main.Backup.describe_report_plan

Main.Backup.describe_restore_job

Main.Backup.disassociate_recovery_point

Main.Backup.export_backup_plan_template

Main.Backup.get_backup_plan

Main.Backup.get_backup_plan_from_json

Main.Backup.get_backup_plan_from_template

Main.Backup.get_backup_selection

Main.Backup.get_backup_vault_access_policy

Main.Backup.get_backup_vault_notifications

Main.Backup.get_recovery_point_restore_metadata

Main.Backup.get_supported_resource_types

Main.Backup.list_backup_jobs

Main.Backup.list_backup_plan_templates

Main.Backup.list_backup_plan_versions

Main.Backup.list_backup_plans

Main.Backup.list_backup_selections

Main.Backup.list_backup_vaults

Main.Backup.list_copy_jobs

Main.Backup.list_frameworks

Main.Backup.list_protected_resources

Main.Backup.list_recovery_points_by_backup_vault

Main.Backup.list_recovery_points_by_resource

Main.Backup.list_report_jobs

Main.Backup.list_report_plans

Main.Backup.list_restore_jobs

Main.Backup.list_tags

Main.Backup.put_backup_vault_access_policy

Main.Backup.put_backup_vault_notifications

Main.Backup.start_backup_job

Main.Backup.start_copy_job

Main.Backup.start_report_job

Main.Backup.start_restore_job

Main.Backup.stop_backup_job

Main.Backup.tag_resource

Main.Backup.untag_resource

Main.Backup.update_backup_plan

Main.Backup.update_framework

Main.Backup.update_global_settings

Main.Backup.update_recovery_point_lifecycle

Main.Backup.update_region_settings

Main.Backup.update_report_plan

Documentation

Main.Backup.create_backup_plan — Method

create_backup_plan(backup_plan)
create_backup_plan(backup_plan, params::Dict{String,<:Any})

Creates a backup plan using a backup plan name and backup rules. A backup plan is a document that contains information that Backup uses to schedule tasks that create recovery points for resources. If you call CreateBackupPlan with a plan that already exists, you receive an AlreadyExistsException exception.

Arguments

backup_plan: Specifies the body of a backup plan. Includes a BackupPlanName and one or more sets of Rules.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"BackupPlanTags": To help organize your resources, you can assign your own metadata to the resources that you create. Each tag is a key-value pair. The specified tags are assigned to all backups created with this plan.

"CreatorRequestId": Identifies the request and allows failed requests to be retried without the risk of running the operation twice. If the request includes a CreatorRequestId that matches an existing backup plan, that plan is returned. This parameter is optional.

Main.Backup.create_backup_selection — Method

create_backup_selection(backup_selection, backup_plan_id)
create_backup_selection(backup_selection, backup_plan_id, params::Dict{String,<:Any})

Creates a JSON document that specifies a set of resources to assign to a backup plan. Resources can be included by specifying patterns for a ListOfTags and selected Resources. For example, consider the following patterns: Resources: "arn:aws:ec2:region:account-id:volume/volume-id" ConditionKey:"department" ConditionValue:"finance" ConditionType:"StringEquals" ConditionKey:"importance" ConditionValue:"critical" ConditionType:"StringEquals" Using these patterns would back up all Amazon Elastic Block Store (Amazon EBS) volumes that are tagged as "department=finance", "importance=critical", in addition to an EBS volume with the specified volume ID. Resources and conditions are additive in that all resources that match the pattern are selected. This shouldn't be confused with a logical AND, where all conditions must match. The matching patterns are logically put together using the OR operator. In other words, all patterns that match are selected for backup.

Arguments

backup_selection: Specifies the body of a request to assign a set of resources to a backup plan.

backup_plan_id: Uniquely identifies the backup plan to be associated with the selection of resources.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"CreatorRequestId": A unique string that identifies the request and allows failed requests to be retried without the risk of running the operation twice.

Main.Backup.create_backup_vault — Method

create_backup_vault(backup_vault_name)
create_backup_vault(backup_vault_name, params::Dict{String,<:Any})

Creates a logical container where backups are stored. A CreateBackupVault request includes a name, optionally one or more resource tags, an encryption key, and a request ID. Do not include sensitive data, such as passport numbers, in the name of a backup vault.

Arguments

backup_vault_name: The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of letters, numbers, and hyphens.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"BackupVaultTags": Metadata that you can assign to help organize the resources that you create. Each tag is a key-value pair.

"CreatorRequestId": A unique string that identifies the request and allows failed requests to be retried without the risk of running the operation twice.

"EncryptionKeyArn": The server-side encryption key that is used to protect your backups; for example, arn:aws:kms:us-west-2:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab.

Main.Backup.create_framework — Method

create_framework(framework_controls, framework_name)
create_framework(framework_controls, framework_name, params::Dict{String,<:Any})

Creates a framework with one or more controls. A framework is a collection of controls that you can use to evaluate your backup practices. By using pre-built customizable controls to define your policies, you can evaluate whether your backup practices comply with your policies. To get insights into the compliance status of your frameworks, you can set up automatic daily reports.

Arguments

framework_controls: A list of the controls that make up the framework. Each control in the list has a name, input parameters, and scope.

framework_name: The unique name of the framework. The name must be between 1 and 256 characters, starting with a letter, and consisting of letters (a-z, A-Z), numbers (0-9), and underscores (_).

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"FrameworkDescription": An optional description of the framework with a maximum of 1,024 characters.

"FrameworkTags": Metadata that you can assign to help organize the frameworks that you create. Each tag is a key-value pair.

"IdempotencyToken": A customer-chosen string that you can use to distinguish between otherwise identical calls to CreateFrameworkInput. Retrying a successful request with the same idempotency token results in a success message with no action taken.

Main.Backup.create_report_plan — Method

create_report_plan(report_delivery_channel, report_plan_name, report_setting)
create_report_plan(report_delivery_channel, report_plan_name, report_setting, params::Dict{String,<:Any})

Creates a report plan. A report plan is a document that contains information about the contents of the report and where Backup will deliver it. If you call CreateReportPlan with a plan that already exists, you receive an AlreadyExistsException exception.

Arguments

report_delivery_channel: A structure that contains information about where and how to deliver your reports, specifically your Amazon S3 bucket name, S3 key prefix, and the formats of your reports.

report_plan_name: The unique name of the report plan. The name must be between 1 and 256 characters, starting with a letter, and consisting of letters (a-z, A-Z), numbers (0-9), and underscores (_).

report_setting: Identifies the report template for the report. Reports are built using a report template. The report templates are: BACKUPJOBREPORT | COPYJOBREPORT | RESTOREJOBREPORT

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"IdempotencyToken": A customer-chosen string that you can use to distinguish between otherwise identical calls to CreateReportPlanInput. Retrying a successful request with the same idempotency token results in a success message with no action taken.

"ReportPlanDescription": An optional description of the report plan with a maximum of 1,024 characters.

"ReportPlanTags": Metadata that you can assign to help organize the frameworks that you create. Each tag is a key-value pair.

Main.Backup.delete_backup_plan — Method

delete_backup_plan(backup_plan_id)
delete_backup_plan(backup_plan_id, params::Dict{String,<:Any})

Deletes a backup plan. A backup plan can only be deleted after all associated selections of resources have been deleted. Deleting a backup plan deletes the current version of a backup plan. Previous versions, if any, will still exist.

Arguments

backup_plan_id: Uniquely identifies a backup plan.

Main.Backup.delete_backup_selection — Method

delete_backup_selection(backup_plan_id, selection_id)
delete_backup_selection(backup_plan_id, selection_id, params::Dict{String,<:Any})

Deletes the resource selection associated with a backup plan that is specified by the SelectionId.

Arguments

backup_plan_id: Uniquely identifies a backup plan.

selection_id: Uniquely identifies the body of a request to assign a set of resources to a backup plan.

Main.Backup.delete_backup_vault — Method

delete_backup_vault(backup_vault_name)
delete_backup_vault(backup_vault_name, params::Dict{String,<:Any})

Deletes the backup vault identified by its name. A vault can be deleted only if it is empty.

Arguments

backup_vault_name: The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

Main.Backup.delete_backup_vault_access_policy — Method

delete_backup_vault_access_policy(backup_vault_name)
delete_backup_vault_access_policy(backup_vault_name, params::Dict{String,<:Any})

Deletes the policy document that manages permissions on a backup vault.

Arguments

backup_vault_name: The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

Main.Backup.delete_backup_vault_notifications — Method

delete_backup_vault_notifications(backup_vault_name)
delete_backup_vault_notifications(backup_vault_name, params::Dict{String,<:Any})

Deletes event notifications for the specified backup vault.

Arguments

backup_vault_name: The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Region where they are created. They consist of lowercase letters, numbers, and hyphens.

Main.Backup.delete_framework — Method

delete_framework(framework_name)
delete_framework(framework_name, params::Dict{String,<:Any})

Deletes the framework specified by a framework name.

Arguments

framework_name: The unique name of a framework.

Main.Backup.delete_recovery_point — Method

delete_recovery_point(backup_vault_name, recovery_point_arn)
delete_recovery_point(backup_vault_name, recovery_point_arn, params::Dict{String,<:Any})

Deletes the recovery point specified by a recovery point ID. If the recovery point ID belongs to a continuous backup, calling this endpoint deletes the existing continuous backup and stops future continuous backup.

Arguments

backup_vault_name: The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

recovery_point_arn: An Amazon Resource Name (ARN) that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

Main.Backup.delete_report_plan — Method

delete_report_plan(report_plan_name)
delete_report_plan(report_plan_name, params::Dict{String,<:Any})

Deletes the report plan specified by a report plan name.

Arguments

report_plan_name: The unique name of a report plan.

Main.Backup.describe_backup_job — Method

describe_backup_job(backup_job_id)
describe_backup_job(backup_job_id, params::Dict{String,<:Any})

Returns backup job details for the specified BackupJobId.

Arguments

backup_job_id: Uniquely identifies a request to Backup to back up a resource.

Main.Backup.describe_backup_vault — Method

describe_backup_vault(backup_vault_name)
describe_backup_vault(backup_vault_name, params::Dict{String,<:Any})

Returns metadata about a backup vault specified by its name.

Arguments

backup_vault_name: The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

Main.Backup.describe_copy_job — Method

describe_copy_job(copy_job_id)
describe_copy_job(copy_job_id, params::Dict{String,<:Any})

Returns metadata associated with creating a copy of a resource.

Arguments

copy_job_id: Uniquely identifies a copy job.

Main.Backup.describe_framework — Method

describe_framework(framework_name)
describe_framework(framework_name, params::Dict{String,<:Any})

Returns the framework details for the specified FrameworkName.

Arguments

framework_name: The unique name of a framework.

Main.Backup.describe_global_settings — Method

describe_global_settings()
describe_global_settings(params::Dict{String,<:Any})

Describes whether the Amazon Web Services account is opted in to cross-account backup. Returns an error if the account is not a member of an Organizations organization. Example: describe-global-settings –region us-west-2

Main.Backup.describe_protected_resource — Method

describe_protected_resource(resource_arn)
describe_protected_resource(resource_arn, params::Dict{String,<:Any})

Returns information about a saved resource, including the last time it was backed up, its Amazon Resource Name (ARN), and the Amazon Web Services service type of the saved resource.

Arguments

resource_arn: An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

Main.Backup.describe_recovery_point — Method

describe_recovery_point(backup_vault_name, recovery_point_arn)
describe_recovery_point(backup_vault_name, recovery_point_arn, params::Dict{String,<:Any})

Returns metadata associated with a recovery point, including ID, status, encryption, and lifecycle.

Arguments

backup_vault_name: The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

recovery_point_arn: An Amazon Resource Name (ARN) that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

Main.Backup.describe_region_settings — Method

describe_region_settings()
describe_region_settings(params::Dict{String,<:Any})

Returns the current service opt-in settings for the Region. If service opt-in is enabled for a service, Backup tries to protect that service's resources in this Region, when the resource is included in an on-demand backup or scheduled backup plan. Otherwise, Backup does not try to protect that service's resources in this Region.

Main.Backup.describe_report_job — Method

describe_report_job(report_job_id)
describe_report_job(report_job_id, params::Dict{String,<:Any})

Returns the details associated with creating a report as specified by its ReportJobId.

Arguments

report_job_id: The identifier of the report job. A unique, randomly generated, Unicode, UTF-8 encoded string that is at most 1,024 bytes long. The report job ID cannot be edited.

Main.Backup.describe_report_plan — Method

describe_report_plan(report_plan_name)
describe_report_plan(report_plan_name, params::Dict{String,<:Any})

Returns a list of all report plans for an Amazon Web Services account and Amazon Web Services Region.

Arguments

report_plan_name: The unique name of a report plan.

Main.Backup.describe_restore_job — Method

describe_restore_job(restore_job_id)
describe_restore_job(restore_job_id, params::Dict{String,<:Any})

Returns metadata associated with a restore job that is specified by a job ID.

Arguments

restore_job_id: Uniquely identifies the job that restores a recovery point.

Main.Backup.disassociate_recovery_point — Method

disassociate_recovery_point(backup_vault_name, recovery_point_arn)
disassociate_recovery_point(backup_vault_name, recovery_point_arn, params::Dict{String,<:Any})

Deletes the specified continuous backup recovery point from Backup and releases control of that continuous backup to the source service, such as Amazon RDS. The source service will continue to create and retain continuous backups using the lifecycle that you specified in your original backup plan. Does not support snapshot backup recovery points.

Arguments

backup_vault_name: The unique name of an Backup vault.

recovery_point_arn: An Amazon Resource Name (ARN) that uniquely identifies an Backup recovery point.

Main.Backup.export_backup_plan_template — Method

export_backup_plan_template(backup_plan_id)
export_backup_plan_template(backup_plan_id, params::Dict{String,<:Any})

Returns the backup plan that is specified by the plan ID as a backup template.

Arguments

backup_plan_id: Uniquely identifies a backup plan.

Main.Backup.get_backup_plan — Method

get_backup_plan(backup_plan_id)
get_backup_plan(backup_plan_id, params::Dict{String,<:Any})

Returns BackupPlan details for the specified BackupPlanId. The details are the body of a backup plan in JSON format, in addition to plan metadata.

Arguments

backup_plan_id: Uniquely identifies a backup plan.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"versionId": Unique, randomly generated, Unicode, UTF-8 encoded strings that are at most 1,024 bytes long. Version IDs cannot be edited.

Main.Backup.get_backup_plan_from_json — Method

get_backup_plan_from_json(backup_plan_template_json)
get_backup_plan_from_json(backup_plan_template_json, params::Dict{String,<:Any})

Returns a valid JSON document specifying a backup plan or an error.

Arguments

backup_plan_template_json: A customer-supplied backup plan document in JSON format.

Main.Backup.get_backup_plan_from_template — Method

get_backup_plan_from_template(template_id)
get_backup_plan_from_template(template_id, params::Dict{String,<:Any})

Returns the template specified by its templateId as a backup plan.

Arguments

template_id: Uniquely identifies a stored backup plan template.

Main.Backup.get_backup_selection — Method

get_backup_selection(backup_plan_id, selection_id)
get_backup_selection(backup_plan_id, selection_id, params::Dict{String,<:Any})

Returns selection metadata and a document in JSON format that specifies a list of resources that are associated with a backup plan.

Arguments

backup_plan_id: Uniquely identifies a backup plan.

selection_id: Uniquely identifies the body of a request to assign a set of resources to a backup plan.

Main.Backup.get_backup_vault_access_policy — Method

get_backup_vault_access_policy(backup_vault_name)
get_backup_vault_access_policy(backup_vault_name, params::Dict{String,<:Any})

Returns the access policy document that is associated with the named backup vault.

Arguments

backup_vault_name: The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

Main.Backup.get_backup_vault_notifications — Method

get_backup_vault_notifications(backup_vault_name)
get_backup_vault_notifications(backup_vault_name, params::Dict{String,<:Any})

Returns event notifications for the specified backup vault.

Arguments

backup_vault_name: The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

Main.Backup.get_recovery_point_restore_metadata — Method

get_recovery_point_restore_metadata(backup_vault_name, recovery_point_arn)
get_recovery_point_restore_metadata(backup_vault_name, recovery_point_arn, params::Dict{String,<:Any})

Returns a set of metadata key-value pairs that were used to create the backup.

Arguments

backup_vault_name: The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

recovery_point_arn: An Amazon Resource Name (ARN) that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

Main.Backup.get_supported_resource_types — Method

get_supported_resource_types()
get_supported_resource_types(params::Dict{String,<:Any})

Returns the Amazon Web Services resource types supported by Backup.

Main.Backup.list_backup_jobs — Method

list_backup_jobs()
list_backup_jobs(params::Dict{String,<:Any})

Returns a list of existing backup jobs for an authenticated account for the last 30 days. For a longer period of time, consider using these monitoring tools.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"accountId": The account ID to list the jobs from. Returns only backup jobs associated with the specified account ID. If used from an Organizations management account, passing * returns all jobs across the organization.

"backupVaultName": Returns only backup jobs that will be stored in the specified backup vault. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

"createdAfter": Returns only backup jobs that were created after the specified date.

"createdBefore": Returns only backup jobs that were created before the specified date.

"maxResults": The maximum number of items to be returned.

"nextToken": The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

"resourceArn": Returns only backup jobs that match the specified resource Amazon Resource Name (ARN).

"resourceType": Returns only backup jobs for the specified resources: DynamoDB for Amazon DynamoDB EBS for Amazon Elastic Block Store EC2 for Amazon Elastic Compute Cloud EFS for Amazon Elastic File System RDS for Amazon Relational Database Service Aurora for Amazon Aurora Storage Gateway for Storage Gateway

"state": Returns only backup jobs that are in the specified state.

Main.Backup.list_backup_plan_templates — Method

list_backup_plan_templates()
list_backup_plan_templates(params::Dict{String,<:Any})

Returns metadata of your saved backup plan templates, including the template ID, name, and the creation and deletion dates.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of items to be returned.

"nextToken": The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

Main.Backup.list_backup_plan_versions — Method

list_backup_plan_versions(backup_plan_id)
list_backup_plan_versions(backup_plan_id, params::Dict{String,<:Any})

Returns version metadata of your backup plans, including Amazon Resource Names (ARNs), backup plan IDs, creation and deletion dates, plan names, and version IDs.

Arguments

backup_plan_id: Uniquely identifies a backup plan.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of items to be returned.

"nextToken": The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

Main.Backup.list_backup_plans — Method

list_backup_plans()
list_backup_plans(params::Dict{String,<:Any})

Returns a list of all active backup plans for an authenticated account. The list contains information such as Amazon Resource Names (ARNs), plan IDs, creation and deletion dates, version IDs, plan names, and creator request IDs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"includeDeleted": A Boolean value with a default value of FALSE that returns deleted backup plans when set to TRUE.

"maxResults": The maximum number of items to be returned.

"nextToken": The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

Main.Backup.list_backup_selections — Method

list_backup_selections(backup_plan_id)
list_backup_selections(backup_plan_id, params::Dict{String,<:Any})

Returns an array containing metadata of the resources associated with the target backup plan.

Arguments

backup_plan_id: Uniquely identifies a backup plan.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of items to be returned.

"nextToken": The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

Main.Backup.list_backup_vaults — Method

list_backup_vaults()
list_backup_vaults(params::Dict{String,<:Any})

Returns a list of recovery point storage containers along with information about them.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of items to be returned.

"nextToken": The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

Main.Backup.list_copy_jobs — Method

list_copy_jobs()
list_copy_jobs(params::Dict{String,<:Any})

Returns metadata about your copy jobs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"accountId": The account ID to list the jobs from. Returns only copy jobs associated with the specified account ID.

"createdAfter": Returns only copy jobs that were created after the specified date.

"createdBefore": Returns only copy jobs that were created before the specified date.

"destinationVaultArn": An Amazon Resource Name (ARN) that uniquely identifies a source backup vault to copy from; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

"maxResults": The maximum number of items to be returned.

"nextToken": The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

"resourceArn": Returns only copy jobs that match the specified resource Amazon Resource Name (ARN).

"resourceType": Returns only backup jobs for the specified resources: DynamoDB for Amazon DynamoDB EBS for Amazon Elastic Block Store EC2 for Amazon Elastic Compute Cloud EFS for Amazon Elastic File System RDS for Amazon Relational Database Service Aurora for Amazon Aurora Storage Gateway for Storage Gateway

"state": Returns only copy jobs that are in the specified state.

Main.Backup.list_frameworks — Method

list_frameworks()
list_frameworks(params::Dict{String,<:Any})

Returns a list of all frameworks for an Amazon Web Services account and Amazon Web Services Region.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The number of desired results from 1 to 1000. Optional. If unspecified, the query will return 1 MB of data.

"NextToken": An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Main.Backup.list_protected_resources — Method

list_protected_resources()
list_protected_resources(params::Dict{String,<:Any})

Returns an array of resources successfully backed up by Backup, including the time the resource was saved, an Amazon Resource Name (ARN) of the resource, and a resource type.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of items to be returned.

"nextToken": The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

Main.Backup.list_recovery_points_by_backup_vault — Method

list_recovery_points_by_backup_vault(backup_vault_name)
list_recovery_points_by_backup_vault(backup_vault_name, params::Dict{String,<:Any})

Returns detailed information about the recovery points stored in a backup vault.

Arguments

backup_vault_name: The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens. Backup vault name might not be available when a supported service creates the backup.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"backupPlanId": Returns only recovery points that match the specified backup plan ID.

"createdAfter": Returns only recovery points that were created after the specified timestamp.

"createdBefore": Returns only recovery points that were created before the specified timestamp.

"maxResults": The maximum number of items to be returned.

"nextToken": The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

"resourceArn": Returns only recovery points that match the specified resource Amazon Resource Name (ARN).

"resourceType": Returns only recovery points that match the specified resource type.

Main.Backup.list_recovery_points_by_resource — Method

list_recovery_points_by_resource(resource_arn)
list_recovery_points_by_resource(resource_arn, params::Dict{String,<:Any})

Returns detailed information about all the recovery points of the type specified by a resource Amazon Resource Name (ARN). For Amazon EFS and Amazon EC2, this action only lists recovery points created by Backup.

Arguments

resource_arn: An ARN that uniquely identifies a resource. The format of the ARN depends on the resource type.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of items to be returned. Amazon RDS requires a value of at least 20.

"nextToken": The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

Main.Backup.list_report_jobs — Method

list_report_jobs()
list_report_jobs(params::Dict{String,<:Any})

Returns details about your report jobs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"CreationAfter": Returns only report jobs that were created after the date and time specified in Unix format and Coordinated Universal Time (UTC). For example, the value 1516925490 represents Friday, January 26, 2018 12:11:30 AM.

"CreationBefore": Returns only report jobs that were created before the date and time specified in Unix format and Coordinated Universal Time (UTC). For example, the value 1516925490 represents Friday, January 26, 2018 12:11:30 AM.

"MaxResults": The number of desired results from 1 to 1000. Optional. If unspecified, the query will return 1 MB of data.

"NextToken": An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

"ReportPlanName": Returns only report jobs with the specified report plan name.

"Status": Returns only report jobs that are in the specified status. The statuses are: CREATED | RUNNING | COMPLETED | FAILED

Main.Backup.list_report_plans — Method

list_report_plans()
list_report_plans(params::Dict{String,<:Any})

Returns a list of your report plans. For detailed information about a single report plan, use DescribeReportPlan.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The number of desired results from 1 to 1000. Optional. If unspecified, the query will return 1 MB of data.

"NextToken": An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Main.Backup.list_restore_jobs — Method

list_restore_jobs()
list_restore_jobs(params::Dict{String,<:Any})

Returns a list of jobs that Backup initiated to restore a saved resource, including details about the recovery process.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"accountId": The account ID to list the jobs from. Returns only restore jobs associated with the specified account ID.

"createdAfter": Returns only restore jobs that were created after the specified date.

"createdBefore": Returns only restore jobs that were created before the specified date.

"maxResults": The maximum number of items to be returned.

"nextToken": The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

"status": Returns only restore jobs associated with the specified job status.

Main.Backup.list_tags — Method

list_tags(resource_arn)
list_tags(resource_arn, params::Dict{String,<:Any})

Returns a list of key-value pairs assigned to a target recovery point, backup plan, or backup vault. ListTags are currently only supported with Amazon EFS backups.

Arguments

resource_arn: An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the type of resource. Valid targets for ListTags are recovery points, backup plans, and backup vaults.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of items to be returned.

"nextToken": The next item following a partial list of returned items. For example, if a request is made to return maxResults number of items, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

Main.Backup.put_backup_vault_access_policy — Method

put_backup_vault_access_policy(backup_vault_name)
put_backup_vault_access_policy(backup_vault_name, params::Dict{String,<:Any})

Sets a resource-based policy that is used to manage access permissions on the target backup vault. Requires a backup vault name and an access policy document in JSON format.

Arguments

backup_vault_name: The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Policy": The backup vault access policy document in JSON format.

Main.Backup.put_backup_vault_notifications — Method

put_backup_vault_notifications(backup_vault_events, snstopic_arn, backup_vault_name)
put_backup_vault_notifications(backup_vault_events, snstopic_arn, backup_vault_name, params::Dict{String,<:Any})

Turns on notifications on a backup vault for the specified topic and events.

Arguments

backup_vault_events: An array of events that indicate the status of jobs to back up resources to the backup vault. The following events are supported: BACKUPJOBSTARTED, BACKUPJOBCOMPLETED, COPYJOBSTARTED, COPYJOBSUCCESSFUL, COPYJOBFAILED, RESTOREJOBSTARTED, RESTOREJOBCOMPLETED, and RECOVERYPOINTMODIFIED. To find failed backup jobs, use BACKUPJOBCOMPLETED and filter using event metadata. Other events in the following list are deprecated.

snstopic_arn: The Amazon Resource Name (ARN) that specifies the topic for a backup vault’s events; for example, arn:aws:sns:us-west-2:111122223333:MyVaultTopic.

backup_vault_name: The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

Main.Backup.start_backup_job — Method

start_backup_job(backup_vault_name, iam_role_arn, resource_arn)
start_backup_job(backup_vault_name, iam_role_arn, resource_arn, params::Dict{String,<:Any})

Starts an on-demand backup job for the specified resource.

Arguments

backup_vault_name: The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

iam_role_arn: Specifies the IAM role ARN used to create the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

resource_arn: An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the resource type.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"BackupOptions": Specifies the backup option for a selected resource. This option is only available for Windows Volume Shadow Copy Service (VSS) backup jobs. Valid values: Set to "WindowsVSS":"enabled" to enable the WindowsVSS backup option and create a Windows VSS backup. Set to "WindowsVSS""disabled" to create a regular backup. The WindowsVSS option is not enabled by default.

"CompleteWindowMinutes": A value in minutes during which a successfully started backup must complete, or else AWS Backup will cancel the job. This value is optional. This value begins counting down from when the backup was scheduled. It does not add additional time for StartWindowMinutes, or if the backup started later than scheduled.

"IdempotencyToken": A customer-chosen string that you can use to distinguish between otherwise identical calls to StartBackupJob. Retrying a successful request with the same idempotency token results in a success message with no action taken.

"Lifecycle": The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup will transition and expire backups automatically according to the lifecycle that you define. Backups transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold. Only Amazon EFS file system backups can be transitioned to cold storage.

"RecoveryPointTags": To help organize your resources, you can assign your own metadata to the resources that you create. Each tag is a key-value pair.

"StartWindowMinutes": A value in minutes after a backup is scheduled before a job will be canceled if it doesn't start successfully. This value is optional, and the default is 8 hours.

Main.Backup.start_copy_job — Method

start_copy_job(destination_backup_vault_arn, iam_role_arn, recovery_point_arn, source_backup_vault_name)
start_copy_job(destination_backup_vault_arn, iam_role_arn, recovery_point_arn, source_backup_vault_name, params::Dict{String,<:Any})

Starts a job to create a one-time copy of the specified resource. Does not support continuous backups.

Arguments

destination_backup_vault_arn: An Amazon Resource Name (ARN) that uniquely identifies a destination backup vault to copy to; for example, arn:aws:backup:us-east-1:123456789012:vault:aBackupVault.

iam_role_arn: Specifies the IAM role ARN used to copy the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

recovery_point_arn: An ARN that uniquely identifies a recovery point to use for the copy job; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

source_backup_vault_name: The name of a logical source container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"IdempotencyToken": A customer-chosen string that you can use to distinguish between otherwise identical calls to StartCopyJob. Retrying a successful request with the same idempotency token results in a success message with no action taken.

"Lifecycle":

Main.Backup.start_report_job — Method

start_report_job(report_plan_name)
start_report_job(report_plan_name, params::Dict{String,<:Any})

Starts an on-demand report job for the specified report plan.

Arguments

report_plan_name: The unique name of a report plan.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"IdempotencyToken": A customer-chosen string that you can use to distinguish between otherwise identical calls to StartReportJobInput. Retrying a successful request with the same idempotency token results in a success message with no action taken.

Main.Backup.start_restore_job — Method

start_restore_job(iam_role_arn, metadata, recovery_point_arn)
start_restore_job(iam_role_arn, metadata, recovery_point_arn, params::Dict{String,<:Any})

Recovers the saved resource identified by an Amazon Resource Name (ARN).

Arguments

iam_role_arn: The Amazon Resource Name (ARN) of the IAM role that Backup uses to create the target recovery point; for example, arn:aws:iam::123456789012:role/S3Access.

metadata: A set of metadata key-value pairs. Contains information, such as a resource name, required to restore a recovery point. You can get configuration metadata about a resource at the time it was backed up by calling GetRecoveryPointRestoreMetadata. However, values in addition to those provided by GetRecoveryPointRestoreMetadata might be required to restore a resource. For example, you might need to provide a new resource name if the original already exists. You need to specify specific metadata to restore an Amazon Elastic File System (Amazon EFS) instance: file-system-id: The ID of the Amazon EFS file system that is backed up by Backup. Returned in GetRecoveryPointRestoreMetadata. Encrypted: A Boolean value that, if true, specifies that the file system is encrypted. If KmsKeyId is specified, Encrypted must be set to true. KmsKeyId: Specifies the Amazon Web Services KMS key that is used to encrypt the restored file system. You can specify a key from another Amazon Web Services account provided that key it is properly shared with your account via Amazon Web Services KMS. PerformanceMode: Specifies the throughput mode of the file system. CreationToken: A user-supplied value that ensures the uniqueness (idempotency) of the request. newFileSystem: A Boolean value that, if true, specifies that the recovery point is restored to a new Amazon EFS file system. ItemsToRestore: An array of one to five strings where each string is a file path. Use ItemsToRestore to restore specific files or directories rather than the entire file system. This parameter is optional. For example, "itemsToRestore":"["/my.test"]".

recovery_point_arn: An ARN that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"IdempotencyToken": A customer-chosen string that you can use to distinguish between otherwise identical calls to StartRestoreJob. Retrying a successful request with the same idempotency token results in a success message with no action taken.

"ResourceType": Starts a job to restore a recovery point for one of the following resources: DynamoDB for Amazon DynamoDB EBS for Amazon Elastic Block Store EC2 for Amazon Elastic Compute Cloud EFS for Amazon Elastic File System RDS for Amazon Relational Database Service Aurora for Amazon Aurora Storage Gateway for Storage Gateway

Main.Backup.stop_backup_job — Method

stop_backup_job(backup_job_id)
stop_backup_job(backup_job_id, params::Dict{String,<:Any})

Attempts to cancel a job to create a one-time backup of a resource.

Arguments

backup_job_id: Uniquely identifies a request to Backup to back up a resource.

Main.Backup.tag_resource — Method

tag_resource(tags, resource_arn)
tag_resource(tags, resource_arn, params::Dict{String,<:Any})

Assigns a set of key-value pairs to a recovery point, backup plan, or backup vault identified by an Amazon Resource Name (ARN).

Arguments

tags: Key-value pairs that are used to help organize your resources. You can assign your own metadata to the resources you create.

resource_arn: An ARN that uniquely identifies a resource. The format of the ARN depends on the type of the tagged resource.

Main.Backup.untag_resource — Method

untag_resource(tag_key_list, resource_arn)
untag_resource(tag_key_list, resource_arn, params::Dict{String,<:Any})

Removes a set of key-value pairs from a recovery point, backup plan, or backup vault identified by an Amazon Resource Name (ARN)

Arguments

tag_key_list: A list of keys to identify which key-value tags to remove from a resource.

resource_arn: An ARN that uniquely identifies a resource. The format of the ARN depends on the type of the tagged resource.

Main.Backup.update_backup_plan — Method

update_backup_plan(backup_plan, backup_plan_id)
update_backup_plan(backup_plan, backup_plan_id, params::Dict{String,<:Any})

Updates an existing backup plan identified by its backupPlanId with the input document in JSON format. The new version is uniquely identified by a VersionId.

Arguments

backup_plan: Specifies the body of a backup plan. Includes a BackupPlanName and one or more sets of Rules.

backup_plan_id: Uniquely identifies a backup plan.

Main.Backup.update_framework — Method

update_framework(framework_name)
update_framework(framework_name, params::Dict{String,<:Any})

Updates an existing framework identified by its FrameworkName with the input document in JSON format.

Arguments

framework_name: The unique name of a framework. This name is between 1 and 256 characters, starting with a letter, and consisting of letters (a-z, A-Z), numbers (0-9), and underscores (_).

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"FrameworkControls": A list of the controls that make up the framework. Each control in the list has a name, input parameters, and scope.

"FrameworkDescription": An optional description of the framework with a maximum 1,024 characters.

"IdempotencyToken": A customer-chosen string that you can use to distinguish between otherwise identical calls to UpdateFrameworkInput. Retrying a successful request with the same idempotency token results in a success message with no action taken.

Main.Backup.update_global_settings — Method

update_global_settings()
update_global_settings(params::Dict{String,<:Any})

Updates whether the Amazon Web Services account is opted in to cross-account backup. Returns an error if the account is not an Organizations management account. Use the DescribeGlobalSettings API to determine the current settings.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"GlobalSettings": A value for isCrossAccountBackupEnabled and a Region. Example: update-global-settings –global-settings isCrossAccountBackupEnabled=false –region us-west-2.

Main.Backup.update_recovery_point_lifecycle — Method

update_recovery_point_lifecycle(backup_vault_name, recovery_point_arn)
update_recovery_point_lifecycle(backup_vault_name, recovery_point_arn, params::Dict{String,<:Any})

Sets the transition lifecycle of a recovery point. The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup transitions and expires backups automatically according to the lifecycle that you define. Backups transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold. Only Amazon EFS file system backups can be transitioned to cold storage. Does not support continuous backups.

Arguments

backup_vault_name: The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the Amazon Web Services Region where they are created. They consist of lowercase letters, numbers, and hyphens.

recovery_point_arn: An Amazon Resource Name (ARN) that uniquely identifies a recovery point; for example, arn:aws:backup:us-east-1:123456789012:recovery-point:1EB3B5E7-9EB0-435A-A80B-108B488B0D45.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Lifecycle": The lifecycle defines when a protected resource is transitioned to cold storage and when it expires. Backup transitions and expires backups automatically according to the lifecycle that you define. Backups transitioned to cold storage must be stored in cold storage for a minimum of 90 days. Therefore, the “expire after days” setting must be 90 days greater than the “transition to cold after days” setting. The “transition to cold after days” setting cannot be changed after a backup has been transitioned to cold.

Main.Backup.update_region_settings — Method

update_region_settings()
update_region_settings(params::Dict{String,<:Any})

Updates the current service opt-in settings for the Region. If service-opt-in is enabled for a service, Backup tries to protect that service's resources in this Region, when the resource is included in an on-demand backup or scheduled backup plan. Otherwise, Backup does not try to protect that service's resources in this Region. Use the DescribeRegionSettings API to determine the resource types that are supported.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ResourceTypeOptInPreference": Updates the list of services along with the opt-in preferences for the Region.

Main.Backup.update_report_plan — Method

update_report_plan(report_plan_name)
update_report_plan(report_plan_name, params::Dict{String,<:Any})

Updates an existing report plan identified by its ReportPlanName with the input document in JSON format.

Arguments

report_plan_name: The unique name of the report plan. This name is between 1 and 256 characters, starting with a letter, and consisting of letters (a-z, A-Z), numbers (0-9), and underscores (_).

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"IdempotencyToken": A customer-chosen string that you can use to distinguish between otherwise identical calls to UpdateReportPlanInput. Retrying a successful request with the same idempotency token results in a success message with no action taken.

"ReportDeliveryChannel": A structure that contains information about where to deliver your reports, specifically your Amazon S3 bucket name, S3 key prefix, and the formats of your reports.

"ReportPlanDescription": An optional description of the report plan with a maximum 1,024 characters.

"ReportSetting": Identifies the report template for the report. Reports are built using a report template. The report templates are: BACKUPJOBREPORT | COPYJOBREPORT | RESTOREJOBREPORT