Backup Gateway

This page documents function available when using the Backup_Gateway module, created with @service Backup_Gateway.

Index

Main.Backup_Gateway.associate_gateway_to_server

Main.Backup_Gateway.create_gateway

Main.Backup_Gateway.delete_gateway

Main.Backup_Gateway.delete_hypervisor

Main.Backup_Gateway.disassociate_gateway_from_server

Main.Backup_Gateway.get_bandwidth_rate_limit_schedule

Main.Backup_Gateway.get_gateway

Main.Backup_Gateway.get_hypervisor

Main.Backup_Gateway.get_hypervisor_property_mappings

Main.Backup_Gateway.get_virtual_machine

Main.Backup_Gateway.import_hypervisor_configuration

Main.Backup_Gateway.list_gateways

Main.Backup_Gateway.list_hypervisors

Main.Backup_Gateway.list_tags_for_resource

Main.Backup_Gateway.list_virtual_machines

Main.Backup_Gateway.put_bandwidth_rate_limit_schedule

Main.Backup_Gateway.put_hypervisor_property_mappings

Main.Backup_Gateway.put_maintenance_start_time

Main.Backup_Gateway.start_virtual_machines_metadata_sync

Main.Backup_Gateway.tag_resource

Main.Backup_Gateway.test_hypervisor_configuration

Main.Backup_Gateway.untag_resource

Main.Backup_Gateway.update_gateway_information

Main.Backup_Gateway.update_gateway_software_now

Main.Backup_Gateway.update_hypervisor

Documentation

Main.Backup_Gateway.associate_gateway_to_server — Method

associate_gateway_to_server(gateway_arn, server_arn)
associate_gateway_to_server(gateway_arn, server_arn, params::Dict{String,<:Any})

Associates a backup gateway with your server. After you complete the association process, you can back up and restore your VMs through the gateway.

Arguments

gateway_arn: The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and Amazon Web Services Region.

server_arn: The Amazon Resource Name (ARN) of the server that hosts your virtual machines.

Main.Backup_Gateway.create_gateway — Method

create_gateway(activation_key, gateway_display_name, gateway_type)
create_gateway(activation_key, gateway_display_name, gateway_type, params::Dict{String,<:Any})

Creates a backup gateway. After you create a gateway, you can associate it with a server using the AssociateGatewayToServer operation.

Arguments

activation_key: The activation key of the created gateway.

gateway_display_name: The display name of the created gateway.

gateway_type: The type of created gateway.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Tags": A list of up to 50 tags to assign to the gateway. Each tag is a key-value pair.

Main.Backup_Gateway.delete_gateway — Method

delete_gateway(gateway_arn)
delete_gateway(gateway_arn, params::Dict{String,<:Any})

Deletes a backup gateway.

Arguments

gateway_arn: The Amazon Resource Name (ARN) of the gateway to delete.

Main.Backup_Gateway.delete_hypervisor — Method

delete_hypervisor(hypervisor_arn)
delete_hypervisor(hypervisor_arn, params::Dict{String,<:Any})

Deletes a hypervisor.

Arguments

hypervisor_arn: The Amazon Resource Name (ARN) of the hypervisor to delete.

Main.Backup_Gateway.disassociate_gateway_from_server — Method

disassociate_gateway_from_server(gateway_arn)
disassociate_gateway_from_server(gateway_arn, params::Dict{String,<:Any})

Disassociates a backup gateway from the specified server. After the disassociation process finishes, the gateway can no longer access the virtual machines on the server.

Arguments

gateway_arn: The Amazon Resource Name (ARN) of the gateway to disassociate.

Main.Backup_Gateway.get_bandwidth_rate_limit_schedule — Method

get_bandwidth_rate_limit_schedule(gateway_arn)
get_bandwidth_rate_limit_schedule(gateway_arn, params::Dict{String,<:Any})

Retrieves the bandwidth rate limit schedule for a specified gateway. By default, gateways do not have bandwidth rate limit schedules, which means no bandwidth rate limiting is in effect. Use this to get a gateway's bandwidth rate limit schedule.

Arguments

gateway_arn: The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and Amazon Web Services Region.

Main.Backup_Gateway.get_gateway — Method

get_gateway(gateway_arn)
get_gateway(gateway_arn, params::Dict{String,<:Any})

By providing the ARN (Amazon Resource Name), this API returns the gateway.

Arguments

gateway_arn: The Amazon Resource Name (ARN) of the gateway.

Main.Backup_Gateway.get_hypervisor — Method

get_hypervisor(hypervisor_arn)
get_hypervisor(hypervisor_arn, params::Dict{String,<:Any})

This action requests information about the specified hypervisor to which the gateway will connect. A hypervisor is hardware, software, or firmware that creates and manages virtual machines, and allocates resources to them.

Arguments

hypervisor_arn: The Amazon Resource Name (ARN) of the hypervisor.

Main.Backup_Gateway.get_hypervisor_property_mappings — Method

get_hypervisor_property_mappings(hypervisor_arn)
get_hypervisor_property_mappings(hypervisor_arn, params::Dict{String,<:Any})

This action retrieves the property mappings for the specified hypervisor. A hypervisor property mapping displays the relationship of entity properties available from the on-premises hypervisor to the properties available in Amazon Web Services.

Arguments

hypervisor_arn: The Amazon Resource Name (ARN) of the hypervisor.

Main.Backup_Gateway.get_virtual_machine — Method

get_virtual_machine(resource_arn)
get_virtual_machine(resource_arn, params::Dict{String,<:Any})

By providing the ARN (Amazon Resource Name), this API returns the virtual machine.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the virtual machine.

Main.Backup_Gateway.import_hypervisor_configuration — Method

import_hypervisor_configuration(host, name)
import_hypervisor_configuration(host, name, params::Dict{String,<:Any})

Connect to a hypervisor by importing its configuration.

Arguments

host: The server host of the hypervisor. This can be either an IP address or a fully-qualified domain name (FQDN).

name: The name of the hypervisor.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"KmsKeyArn": The Key Management Service for the hypervisor.

"Password": The password for the hypervisor.

"Tags": The tags of the hypervisor configuration to import.

"Username": The username for the hypervisor.

Main.Backup_Gateway.list_gateways — Method

list_gateways()
list_gateways(params::Dict{String,<:Any})

Lists backup gateways owned by an Amazon Web Services account in an Amazon Web Services Region. The returned list is ordered by gateway Amazon Resource Name (ARN).

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of gateways to list.

"NextToken": The next item following a partial list of returned resources. For example, if a request is made to return MaxResults number of resources, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

Main.Backup_Gateway.list_hypervisors — Method

list_hypervisors()
list_hypervisors(params::Dict{String,<:Any})

Lists your hypervisors.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of hypervisors to list.

"NextToken": The next item following a partial list of returned resources. For example, if a request is made to return maxResults number of resources, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

Main.Backup_Gateway.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists the tags applied to the resource identified by its Amazon Resource Name (ARN).

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource's tags to list.

Main.Backup_Gateway.list_virtual_machines — Method

list_virtual_machines()
list_virtual_machines(params::Dict{String,<:Any})

Lists your virtual machines.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"HypervisorArn": The Amazon Resource Name (ARN) of the hypervisor connected to your virtual machine.

"MaxResults": The maximum number of virtual machines to list.

"NextToken": The next item following a partial list of returned resources. For example, if a request is made to return maxResults number of resources, NextToken allows you to return more items in your list starting at the location pointed to by the next token.

Main.Backup_Gateway.put_bandwidth_rate_limit_schedule — Method

put_bandwidth_rate_limit_schedule(bandwidth_rate_limit_intervals, gateway_arn)
put_bandwidth_rate_limit_schedule(bandwidth_rate_limit_intervals, gateway_arn, params::Dict{String,<:Any})

This action sets the bandwidth rate limit schedule for a specified gateway. By default, gateways do not have a bandwidth rate limit schedule, which means no bandwidth rate limiting is in effect. Use this to initiate a gateway's bandwidth rate limit schedule.

Arguments

bandwidth_rate_limit_intervals: An array containing bandwidth rate limit schedule intervals for a gateway. When no bandwidth rate limit intervals have been scheduled, the array is empty.

gateway_arn: The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and Amazon Web Services Region.

Main.Backup_Gateway.put_hypervisor_property_mappings — Method

put_hypervisor_property_mappings(hypervisor_arn, iam_role_arn, vmware_to_aws_tag_mappings)
put_hypervisor_property_mappings(hypervisor_arn, iam_role_arn, vmware_to_aws_tag_mappings, params::Dict{String,<:Any})

This action sets the property mappings for the specified hypervisor. A hypervisor property mapping displays the relationship of entity properties available from the on-premises hypervisor to the properties available in Amazon Web Services.

Arguments

hypervisor_arn: The Amazon Resource Name (ARN) of the hypervisor.

iam_role_arn: The Amazon Resource Name (ARN) of the IAM role.

vmware_to_aws_tag_mappings: This action requests the mappings of on-premises VMware tags to the Amazon Web Services tags.

Main.Backup_Gateway.put_maintenance_start_time — Method

put_maintenance_start_time(gateway_arn, hour_of_day, minute_of_hour)
put_maintenance_start_time(gateway_arn, hour_of_day, minute_of_hour, params::Dict{String,<:Any})

Set the maintenance start time for a gateway.

Arguments

gateway_arn: The Amazon Resource Name (ARN) for the gateway, used to specify its maintenance start time.

hour_of_day: The hour of the day to start maintenance on a gateway.

minute_of_hour: The minute of the hour to start maintenance on a gateway.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DayOfMonth": The day of the month start maintenance on a gateway. Valid values range from Sunday to Saturday.

"DayOfWeek": The day of the week to start maintenance on a gateway.

Main.Backup_Gateway.start_virtual_machines_metadata_sync — Method

start_virtual_machines_metadata_sync(hypervisor_arn)
start_virtual_machines_metadata_sync(hypervisor_arn, params::Dict{String,<:Any})

This action sends a request to sync metadata across the specified virtual machines.

Arguments

hypervisor_arn: The Amazon Resource Name (ARN) of the hypervisor.

Main.Backup_Gateway.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Tag the resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource to tag.

tags: A list of tags to assign to the resource.

Main.Backup_Gateway.test_hypervisor_configuration — Method

test_hypervisor_configuration(gateway_arn, host)
test_hypervisor_configuration(gateway_arn, host, params::Dict{String,<:Any})

Tests your hypervisor configuration to validate that backup gateway can connect with the hypervisor and its resources.

Arguments

gateway_arn: The Amazon Resource Name (ARN) of the gateway to the hypervisor to test.

host: The server host of the hypervisor. This can be either an IP address or a fully-qualified domain name (FQDN).

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Password": The password for the hypervisor.

"Username": The username for the hypervisor.

Main.Backup_Gateway.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes tags from the resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource from which to remove tags.

tag_keys: The list of tag keys specifying which tags to remove.

Main.Backup_Gateway.update_gateway_information — Method

update_gateway_information(gateway_arn)
update_gateway_information(gateway_arn, params::Dict{String,<:Any})

Updates a gateway's name. Specify which gateway to update using the Amazon Resource Name (ARN) of the gateway in your request.

Arguments

gateway_arn: The Amazon Resource Name (ARN) of the gateway to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"GatewayDisplayName": The updated display name of the gateway.

Main.Backup_Gateway.update_gateway_software_now — Method

update_gateway_software_now(gateway_arn)
update_gateway_software_now(gateway_arn, params::Dict{String,<:Any})

Updates the gateway virtual machine (VM) software. The request immediately triggers the software update. When you make this request, you get a 200 OK success response immediately. However, it might take some time for the update to complete.

Arguments

gateway_arn: The Amazon Resource Name (ARN) of the gateway to be updated.

Main.Backup_Gateway.update_hypervisor — Method

update_hypervisor(hypervisor_arn)
update_hypervisor(hypervisor_arn, params::Dict{String,<:Any})

Updates a hypervisor metadata, including its host, username, and password. Specify which hypervisor to update using the Amazon Resource Name (ARN) of the hypervisor in your request.

Arguments

hypervisor_arn: The Amazon Resource Name (ARN) of the hypervisor to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Host": The updated host of the hypervisor. This can be either an IP address or a fully-qualified domain name (FQDN).

"LogGroupArn": The Amazon Resource Name (ARN) of the group of gateways within the requested log.

"Name": The updated name for the hypervisor

"Password": The updated password for the hypervisor.

"Username": The updated username for the hypervisor.