Chatbot

This page documents function available when using the Chatbot module, created with @service Chatbot.

Index

Main.Chatbot.associate_to_configuration

Main.Chatbot.create_chime_webhook_configuration

Main.Chatbot.create_custom_action

Main.Chatbot.create_microsoft_teams_channel_configuration

Main.Chatbot.create_slack_channel_configuration

Main.Chatbot.delete_chime_webhook_configuration

Main.Chatbot.delete_custom_action

Main.Chatbot.delete_microsoft_teams_channel_configuration

Main.Chatbot.delete_microsoft_teams_configured_team

Main.Chatbot.delete_microsoft_teams_user_identity

Main.Chatbot.delete_slack_channel_configuration

Main.Chatbot.delete_slack_user_identity

Main.Chatbot.delete_slack_workspace_authorization

Main.Chatbot.describe_chime_webhook_configurations

Main.Chatbot.describe_slack_channel_configurations

Main.Chatbot.describe_slack_user_identities

Main.Chatbot.describe_slack_workspaces

Main.Chatbot.disassociate_from_configuration

Main.Chatbot.get_account_preferences

Main.Chatbot.get_custom_action

Main.Chatbot.get_microsoft_teams_channel_configuration

Main.Chatbot.list_associations

Main.Chatbot.list_custom_actions

Main.Chatbot.list_microsoft_teams_channel_configurations

Main.Chatbot.list_microsoft_teams_configured_teams

Main.Chatbot.list_microsoft_teams_user_identities

Main.Chatbot.list_tags_for_resource

Main.Chatbot.tag_resource

Main.Chatbot.untag_resource

Main.Chatbot.update_account_preferences

Main.Chatbot.update_chime_webhook_configuration

Main.Chatbot.update_custom_action

Main.Chatbot.update_microsoft_teams_channel_configuration

Main.Chatbot.update_slack_channel_configuration

Documentation

Main.Chatbot.associate_to_configuration — Function

associate_to_configuration(chat_configuration, resource)
associate_to_configuration(chat_configuration, resource, params::Dict{String,<:Any})

Links a resource (for example, a custom action) to a channel configuration.

Arguments

chat_configuration: The channel configuration to associate with the resource.

resource: The resource Amazon Resource Name (ARN) to link.

Main.Chatbot.create_chime_webhook_configuration — Function

create_chime_webhook_configuration(configuration_name, iam_role_arn, sns_topic_arns, webhook_description, webhook_url)
create_chime_webhook_configuration(configuration_name, iam_role_arn, sns_topic_arns, webhook_description, webhook_url, params::Dict{String,<:Any})

Creates an AWS Chatbot configuration for Amazon Chime.

Arguments

configuration_name: The name of the configuration.

iam_role_arn: A user-defined role that AWS Chatbot assumes. This is not the service- linked role.

For more information, see IAM policies for AWS Chatbot in the AWS Chatbot Administrator Guide.

sns_topic_arns: The Amazon Resource Names (ARNs) of the SNS topics that deliver notifications to AWS Chatbot.

webhook_description: A description of the webhook. We recommend using the convention RoomName/WebhookName.

For more information, see Tutorial: Get started with Amazon Chime in the AWS Chatbot Administrator Guide.

webhook_url: The URL for the Amazon Chime webhook.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"LoggingLevel": Logging levels include ERROR, INFO, or NONE.

"Tags": A map of tags assigned to a resource. A tag is a string-to-string map of key- value pairs.

Main.Chatbot.create_custom_action — Function

create_custom_action(action_name, definition)
create_custom_action(action_name, definition, params::Dict{String,<:Any})

Creates a custom action that can be invoked as an alias or as a button on a notification.

Arguments

action_name: The name of the custom action. This name is included in the Amazon Resource Name (ARN).

definition: The definition of the command to run when invoked as an alias or as an action button.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AliasName": The name used to invoke this action in a chat channel. For example, @aws run my-alias.

"Attachments": Defines when this custom action button should be attached to a notification.

"ClientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, subsequent retries with the same client token returns the result from the original successful request.

If you do not specify a client token, one is automatically generated by the SDK.

"Tags": A map of tags assigned to a resource. A tag is a string-to-string map of key- value pairs.

Main.Chatbot.create_microsoft_teams_channel_configuration — Function

create_microsoft_teams_channel_configuration(channel_id, configuration_name, iam_role_arn, team_id, tenant_id)
create_microsoft_teams_channel_configuration(channel_id, configuration_name, iam_role_arn, team_id, tenant_id, params::Dict{String,<:Any})

Creates an AWS Chatbot configuration for Microsoft Teams.

Arguments

channel_id: The ID of the Microsoft Teams channel.

configuration_name: The name of the configuration.

iam_role_arn: A user-defined role that AWS Chatbot assumes. This is not the service- linked role.

For more information, see IAM policies for AWS Chatbot in the AWS Chatbot Administrator Guide.

team_id: The ID of the Microsoft Teams authorized with AWS Chatbot.

To get the team ID, you must perform the initial authorization flow with Microsoft Teams in the AWS Chatbot console. Then you can copy and paste the team ID from the console. For more information, see Step 1: Configure a Microsoft Teams client in the AWS Chatbot Administrator Guide.

tenant_id: The ID of the Microsoft Teams tenant.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ChannelName": The name of the Microsoft Teams channel.

"GuardrailPolicyArns": The list of IAM policy ARNs that are applied as channel guardrails. The AWS managed AdministratorAccess policy is applied by default if this is not set.

"LoggingLevel": Logging levels include ERROR, INFO, or NONE.

"SnsTopicArns": The Amazon Resource Names (ARNs) of the SNS topics that deliver notifications to AWS Chatbot.

"Tags": A map of tags assigned to a resource. A tag is a string-to-string map of key- value pairs.

"TeamName": The name of the Microsoft Teams Team.

"UserAuthorizationRequired": Enables use of a user role requirement in your chat configuration.

Main.Chatbot.create_slack_channel_configuration — Function

create_slack_channel_configuration(configuration_name, iam_role_arn, slack_channel_id, slack_team_id)
create_slack_channel_configuration(configuration_name, iam_role_arn, slack_channel_id, slack_team_id, params::Dict{String,<:Any})

Creates an AWS Chatbot confugration for Slack.

Arguments

configuration_name: The name of the configuration.

iam_role_arn: A user-defined role that AWS Chatbot assumes. This is not the service- linked role.

For more information, see IAM policies for AWS Chatbot in the AWS Chatbot Administrator Guide.

slack_channel_id: The ID of the Slack channel.

To get this ID, open Slack, right click on the channel name in the left pane, then choose Copy Link. The channel ID is the 9-character string at the end of the URL. For example, ABCBBLZZZ.

slack_team_id: The ID of the Slack workspace authorized with AWS Chatbot.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"GuardrailPolicyArns": The list of IAM policy ARNs that are applied as channel guardrails. The AWS managed AdministratorAccess policy is applied by default if this is not set.

"LoggingLevel": Logging levels include ERROR, INFO, or NONE.

"SlackChannelName": The name of the Slack channel.

"SnsTopicArns": The Amazon Resource Names (ARNs) of the SNS topics that deliver notifications to AWS Chatbot.

"Tags": A map of tags assigned to a resource. A tag is a string-to-string map of key- value pairs.

"UserAuthorizationRequired": Enables use of a user role requirement in your chat configuration.

Main.Chatbot.delete_chime_webhook_configuration — Function

delete_chime_webhook_configuration(chat_configuration_arn)
delete_chime_webhook_configuration(chat_configuration_arn, params::Dict{String,<:Any})

Deletes a Amazon Chime webhook configuration for AWS Chatbot.

Arguments

chat_configuration_arn: The Amazon Resource Name (ARN) of the ChimeWebhookConfiguration to delete.

Main.Chatbot.delete_custom_action — Function

delete_custom_action(custom_action_arn)
delete_custom_action(custom_action_arn, params::Dict{String,<:Any})

Deletes a custom action.

Arguments

custom_action_arn: The fully defined ARN of the custom action.

Main.Chatbot.delete_microsoft_teams_channel_configuration — Function

delete_microsoft_teams_channel_configuration(chat_configuration_arn)
delete_microsoft_teams_channel_configuration(chat_configuration_arn, params::Dict{String,<:Any})

Deletes a Microsoft Teams channel configuration for AWS Chatbot

Arguments

chat_configuration_arn: The Amazon Resource Name (ARN) of the MicrosoftTeamsChannelConfiguration associated with the user identity to delete.

Main.Chatbot.delete_microsoft_teams_configured_team — Function

delete_microsoft_teams_configured_team(team_id)
delete_microsoft_teams_configured_team(team_id, params::Dict{String,<:Any})

Deletes the Microsoft Teams team authorization allowing for channels to be configured in that Microsoft Teams team. Note that the Microsoft Teams team must have no channels configured to remove it.

Arguments

team_id: The ID of the Microsoft Teams team authorized with AWS Chatbot.

To get the team ID, you must perform the initial authorization flow with Microsoft Teams in the AWS Chatbot console. Then you can copy and paste the team ID from the console. For more information, see Step 1: Configure a Microsoft Teams client in the AWS Chatbot Administrator Guide.

Main.Chatbot.delete_microsoft_teams_user_identity — Function

delete_microsoft_teams_user_identity(chat_configuration_arn, user_id)
delete_microsoft_teams_user_identity(chat_configuration_arn, user_id, params::Dict{String,<:Any})

Identifes a user level permission for a channel configuration.

Arguments

chat_configuration_arn: The ARN of the MicrosoftTeamsChannelConfiguration associated with the user identity to delete.

user_id: The Microsoft Teams user ID.

Main.Chatbot.delete_slack_channel_configuration — Function

delete_slack_channel_configuration(chat_configuration_arn)
delete_slack_channel_configuration(chat_configuration_arn, params::Dict{String,<:Any})

Deletes a Slack channel configuration for AWS Chatbot

Arguments

chat_configuration_arn: The Amazon Resource Name (ARN) of the SlackChannelConfiguration to delete.

Main.Chatbot.delete_slack_user_identity — Function

delete_slack_user_identity(chat_configuration_arn, slack_team_id, slack_user_id)
delete_slack_user_identity(chat_configuration_arn, slack_team_id, slack_user_id, params::Dict{String,<:Any})

Deletes a user level permission for a Slack channel configuration.

Arguments

chat_configuration_arn: The ARN of the SlackChannelConfiguration associated with the user identity to delete.

slack_team_id: The ID of the Slack workspace authorized with AWS Chatbot.

slack_user_id: The ID of the user in Slack

Main.Chatbot.delete_slack_workspace_authorization — Function

delete_slack_workspace_authorization(slack_team_id)
delete_slack_workspace_authorization(slack_team_id, params::Dict{String,<:Any})

Deletes the Slack workspace authorization that allows channels to be configured in that workspace. This requires all configured channels in the workspace to be deleted.

Arguments

slack_team_id: The ID of the Slack workspace authorized with AWS Chatbot.

Main.Chatbot.describe_chime_webhook_configurations — Function

describe_chime_webhook_configurations()
describe_chime_webhook_configurations(params::Dict{String,<:Any})

Lists Amazon Chime webhook configurations optionally filtered by ChatConfigurationArn

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ChatConfigurationArn": An optional Amazon Resource Name (ARN) of a ChimeWebhookConfiguration to describe.

"MaxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"NextToken": An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

Main.Chatbot.describe_slack_channel_configurations — Function

describe_slack_channel_configurations()
describe_slack_channel_configurations(params::Dict{String,<:Any})

Lists Slack channel configurations optionally filtered by ChatConfigurationArn

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ChatConfigurationArn": An optional Amazon Resource Name (ARN) of a SlackChannelConfiguration to describe.

"MaxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"NextToken": An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

Main.Chatbot.describe_slack_user_identities — Function

describe_slack_user_identities()
describe_slack_user_identities(params::Dict{String,<:Any})

Lists all Slack user identities with a mapped role.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ChatConfigurationArn": The Amazon Resource Name (ARN) of the SlackChannelConfiguration associated with the user identities to describe.

"MaxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"NextToken": An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

Main.Chatbot.describe_slack_workspaces — Function

describe_slack_workspaces()
describe_slack_workspaces(params::Dict{String,<:Any})

List all authorized Slack workspaces connected to the AWS Account onboarded with AWS Chatbot.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"NextToken": An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

Main.Chatbot.disassociate_from_configuration — Function

disassociate_from_configuration(chat_configuration, resource)
disassociate_from_configuration(chat_configuration, resource, params::Dict{String,<:Any})

Unlink a resource, for example a custom action, from a channel configuration.

Arguments

chat_configuration: The channel configuration the resource is being disassociated from.

resource: The resource (for example, a custom action) Amazon Resource Name (ARN) to unlink.

Main.Chatbot.get_account_preferences — Function

get_account_preferences()
get_account_preferences(params::Dict{String,<:Any})

Returns AWS Chatbot account preferences.

Main.Chatbot.get_custom_action — Function

get_custom_action(custom_action_arn)
get_custom_action(custom_action_arn, params::Dict{String,<:Any})

Returns a custom action.

Arguments

custom_action_arn: Returns the fully defined Amazon Resource Name (ARN) of the custom action.

Main.Chatbot.get_microsoft_teams_channel_configuration — Function

get_microsoft_teams_channel_configuration(chat_configuration_arn)
get_microsoft_teams_channel_configuration(chat_configuration_arn, params::Dict{String,<:Any})

Returns a Microsoft Teams channel configuration in an AWS account.

Arguments

chat_configuration_arn: The Amazon Resource Name (ARN) of the MicrosoftTeamsChannelConfiguration to retrieve.

Main.Chatbot.list_associations — Function

list_associations(chat_configuration)
list_associations(chat_configuration, params::Dict{String,<:Any})

Lists resources associated with a channel configuration.

Arguments

chat_configuration: The channel configuration to list associations for.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"NextToken": An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

Main.Chatbot.list_custom_actions — Function

list_custom_actions()
list_custom_actions(params::Dict{String,<:Any})

Lists custom actions defined in this account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"NextToken": An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

Main.Chatbot.list_microsoft_teams_channel_configurations — Function

list_microsoft_teams_channel_configurations()
list_microsoft_teams_channel_configurations(params::Dict{String,<:Any})

Lists all AWS Chatbot Microsoft Teams channel configurations in an AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"NextToken": An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

"TeamId": The ID of the Microsoft Teams authorized with AWS Chatbot.

To get the team ID, you must perform the initial authorization flow with Microsoft Teams in the AWS Chatbot console. Then you can copy and paste the team ID from the console. For more information, see Step 1: Configure a Microsoft Teams client in the AWS Chatbot Administrator Guide.

Main.Chatbot.list_microsoft_teams_configured_teams — Function

list_microsoft_teams_configured_teams()
list_microsoft_teams_configured_teams(params::Dict{String,<:Any})

Lists all authorized Microsoft Teams for an AWS Account

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"NextToken": An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

Main.Chatbot.list_microsoft_teams_user_identities — Function

list_microsoft_teams_user_identities()
list_microsoft_teams_user_identities(params::Dict{String,<:Any})

A list all Microsoft Teams user identities with a mapped role.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ChatConfigurationArn": The Amazon Resource Name (ARN) of the MicrosoftTeamsChannelConfiguration associated with the user identities to list.

"MaxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"NextToken": An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

Main.Chatbot.list_tags_for_resource — Function

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists all of the tags associated with the Amazon Resource Name (ARN) that you specify. The resource can be a user, server, or role.

Arguments

resource_arn: The ARN of the resource to list tags for.

Main.Chatbot.tag_resource — Function

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Attaches a key-value pair to a resource, as identified by its Amazon Resource Name (ARN). Resources are users, servers, roles, and other entities.

Arguments

resource_arn: The ARN of the configuration.

tags: A list of tags to apply to the configuration.

Main.Chatbot.untag_resource — Function

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Detaches a key-value pair from a resource, as identified by its Amazon Resource Name (ARN). Resources are users, servers, roles, and other entities.

Arguments

resource_arn: The value of the resource that will have the tag removed. An Amazon Resource Name (ARN) is an identifier for a specific AWS resource, such as a server, user, or role.

tag_keys: TagKeys are key-value pairs assigned to ARNs that can be used to group and search for resources by type. This metadata can be attached to resources for any purpose.

Main.Chatbot.update_account_preferences — Function

update_account_preferences()
update_account_preferences(params::Dict{String,<:Any})

Updates AWS Chatbot account preferences.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"TrainingDataCollectionEnabled": Turns on training data collection.

This helps improve the AWS Chatbot experience by allowing AWS Chatbot to store and use your customer information, such as AWS Chatbot configurations, notifications, user inputs, AWS Chatbot generated responses, and interaction data. This data helps us to continuously improve and develop Artificial Intelligence (AI) technologies. Your data is not shared with any third parties and is protected using sophisticated controls to prevent unauthorized access and misuse. AWS Chatbot does not store or use interactions in chat channels with Amazon Q for training AI technologies for AWS Chatbot.

"UserAuthorizationRequired": Enables use of a user role requirement in your chat configuration.

Main.Chatbot.update_chime_webhook_configuration — Function

update_chime_webhook_configuration(chat_configuration_arn)
update_chime_webhook_configuration(chat_configuration_arn, params::Dict{String,<:Any})

Updates a Amazon Chime webhook configuration.

Arguments

chat_configuration_arn: The Amazon Resource Name (ARN) of the ChimeWebhookConfiguration to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"IamRoleArn": A user-defined role that AWS Chatbot assumes. This is not the service- linked role.

For more information, see IAM policies for AWS Chatbot in the AWS Chatbot Administrator Guide.

"LoggingLevel": Logging levels include ERROR, INFO, or NONE.

"SnsTopicArns": The ARNs of the SNS topics that deliver notifications to AWS Chatbot.

"WebhookDescription": A description of the webhook. We recommend using the convention RoomName/WebhookName.

For more information, see Tutorial: Get started with Amazon Chime in the AWS Chatbot Administrator Guide.

"WebhookUrl": The URL for the Amazon Chime webhook.

Main.Chatbot.update_custom_action — Function

update_custom_action(custom_action_arn, definition)
update_custom_action(custom_action_arn, definition, params::Dict{String,<:Any})

Updates a custom action.

Arguments

custom_action_arn: The fully defined Amazon Resource Name (ARN) of the custom action.

definition: The definition of the command to run when invoked as an alias or as an action button.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AliasName": The name used to invoke this action in the chat channel. For example, @aws run my-alias.

"Attachments": Defines when this custom action button should be attached to a notification.

Main.Chatbot.update_microsoft_teams_channel_configuration — Function

update_microsoft_teams_channel_configuration(channel_id, chat_configuration_arn)
update_microsoft_teams_channel_configuration(channel_id, chat_configuration_arn, params::Dict{String,<:Any})

Updates an Microsoft Teams channel configuration.

Arguments

channel_id: The ID of the Microsoft Teams channel.

chat_configuration_arn: The Amazon Resource Name (ARN) of the TeamsChannelConfiguration to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ChannelName": The name of the Microsoft Teams channel.

"GuardrailPolicyArns": The list of IAM policy ARNs that are applied as channel guardrails. The AWS managed AdministratorAccess policy is applied by default if this is not set.

"IamRoleArn": A user-defined role that AWS Chatbot assumes. This is not the service- linked role.

For more information, see IAM policies for AWS Chatbot in the AWS Chatbot Administrator Guide.

"LoggingLevel": Logging levels include ERROR, INFO, or NONE.

"SnsTopicArns": The Amazon Resource Names (ARNs) of the SNS topics that deliver notifications to AWS Chatbot.

"UserAuthorizationRequired": Enables use of a user role requirement in your chat configuration.

Main.Chatbot.update_slack_channel_configuration — Function

update_slack_channel_configuration(chat_configuration_arn, slack_channel_id)
update_slack_channel_configuration(chat_configuration_arn, slack_channel_id, params::Dict{String,<:Any})

Updates a Slack channel configuration.

Arguments

chat_configuration_arn: The Amazon Resource Name (ARN) of the SlackChannelConfiguration to update.

slack_channel_id: The ID of the Slack channel.

To get this ID, open Slack, right click on the channel name in the left pane, then choose Copy Link. The channel ID is the 9-character string at the end of the URL. For example, ABCBBLZZZ.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"GuardrailPolicyArns": The list of IAM policy ARNs that are applied as channel guardrails. The AWS managed AdministratorAccess policy is applied by default if this is not set.

"IamRoleArn": A user-defined role that AWS Chatbot assumes. This is not the service- linked role.

For more information, see IAM policies for AWS Chatbot in the AWS Chatbot Administrator Guide.

"LoggingLevel": Logging levels include ERROR, INFO, or NONE.

"SlackChannelName": The name of the Slack channel.

"SnsTopicArns": The Amazon Resource Names (ARNs) of the SNS topics that deliver notifications to AWS Chatbot.

"UserAuthorizationRequired": Enables use of a user role requirement in your chat configuration.