Chime

This page documents function available when using the Chime module, created with @service Chime.

Index

Main.Chime.associate_phone_number_with_user

Main.Chime.associate_signin_delegate_groups_with_account

Main.Chime.batch_create_room_membership

Main.Chime.batch_delete_phone_number

Main.Chime.batch_suspend_user

Main.Chime.batch_unsuspend_user

Main.Chime.batch_update_phone_number

Main.Chime.batch_update_user

Main.Chime.create_account

Main.Chime.create_bot

Main.Chime.create_meeting_dial_out

Main.Chime.create_phone_number_order

Main.Chime.create_room

Main.Chime.create_room_membership

Main.Chime.create_user

Main.Chime.delete_account

Main.Chime.delete_events_configuration

Main.Chime.delete_phone_number

Main.Chime.delete_room

Main.Chime.delete_room_membership

Main.Chime.disassociate_phone_number_from_user

Main.Chime.disassociate_signin_delegate_groups_from_account

Main.Chime.get_account

Main.Chime.get_account_settings

Main.Chime.get_bot

Main.Chime.get_events_configuration

Main.Chime.get_global_settings

Main.Chime.get_phone_number

Main.Chime.get_phone_number_order

Main.Chime.get_phone_number_settings

Main.Chime.get_retention_settings

Main.Chime.get_room

Main.Chime.get_user

Main.Chime.get_user_settings

Main.Chime.invite_users

Main.Chime.list_accounts

Main.Chime.list_bots

Main.Chime.list_phone_number_orders

Main.Chime.list_phone_numbers

Main.Chime.list_room_memberships

Main.Chime.list_rooms

Main.Chime.list_supported_phone_number_countries

Main.Chime.list_users

Main.Chime.logout_user

Main.Chime.put_events_configuration

Main.Chime.put_retention_settings

Main.Chime.redact_conversation_message

Main.Chime.redact_room_message

Main.Chime.regenerate_security_token

Main.Chime.reset_personal_pin

Main.Chime.restore_phone_number

Main.Chime.search_available_phone_numbers

Main.Chime.update_account

Main.Chime.update_account_settings

Main.Chime.update_bot

Main.Chime.update_global_settings

Main.Chime.update_phone_number

Main.Chime.update_phone_number_settings

Main.Chime.update_room

Main.Chime.update_room_membership

Main.Chime.update_user

Main.Chime.update_user_settings

Documentation

Main.Chime.associate_phone_number_with_user — Function

associate_phone_number_with_user(account_id, e164_phone_number, user_id)
associate_phone_number_with_user(account_id, e164_phone_number, user_id, params::Dict{String,<:Any})

Associates a phone number with the specified Amazon Chime user.

Arguments

account_id: The Amazon Chime account ID.

e164_phone_number: The phone number, in E.164 format.

user_id: The user ID.

Main.Chime.associate_signin_delegate_groups_with_account — Function

associate_signin_delegate_groups_with_account(account_id, signin_delegate_groups)
associate_signin_delegate_groups_with_account(account_id, signin_delegate_groups, params::Dict{String,<:Any})

Associates the specified sign-in delegate groups with the specified Amazon Chime account.

Arguments

account_id: The Amazon Chime account ID.

signin_delegate_groups: The sign-in delegate groups.

Main.Chime.batch_create_room_membership — Function

batch_create_room_membership(account_id, membership_item_list, room_id)
batch_create_room_membership(account_id, membership_item_list, room_id, params::Dict{String,<:Any})

Adds up to 50 members to a chat room in an Amazon Chime Enterprise account. Members can be users or bots. The member role designates whether the member is a chat room administrator or a general chat room member.

Arguments

account_id: The Amazon Chime account ID.

membership_item_list: The list of membership items.

room_id: The room ID.

Main.Chime.batch_delete_phone_number — Function

batch_delete_phone_number(phone_number_ids)
batch_delete_phone_number(phone_number_ids, params::Dict{String,<:Any})

Moves phone numbers into the Deletion queue. Phone numbers must be disassociated from any users or Amazon Chime Voice Connectors before they can be deleted.

Phone numbers remain in the Deletion queue for 7 days before they are deleted permanently.

Arguments

phone_number_ids: List of phone number IDs.

Main.Chime.batch_suspend_user — Function

batch_suspend_user(account_id, user_id_list)
batch_suspend_user(account_id, user_id_list, params::Dict{String,<:Any})

Suspends up to 50 users from a Team or EnterpriseLWA Amazon Chime account. For more information about different account types, see Managing Your Amazon Chime Accounts in the Amazon Chime Administration Guide.

Users suspended from a Team account are disassociated from the account,but they can continue to use Amazon Chime as free users. To remove the suspension from suspended Team account users, invite them to the Team account again. You can use the invite_users action to do so.

Users suspended from an EnterpriseLWA account are immediately signed out of Amazon Chime and can no longer sign in. To remove the suspension from suspended EnterpriseLWA account users, use the batch_unsuspend_user action.

To sign out users without suspending them, use the logout_user action.

Arguments

account_id: The Amazon Chime account ID.

user_id_list: The request containing the user IDs to suspend.

Main.Chime.batch_unsuspend_user — Function

batch_unsuspend_user(account_id, user_id_list)
batch_unsuspend_user(account_id, user_id_list, params::Dict{String,<:Any})

Removes the suspension from up to 50 previously suspended users for the specified Amazon Chime EnterpriseLWA account. Only users on EnterpriseLWA accounts can be unsuspended using this action. For more information about different account types, see Managing Your Amazon Chime Accounts in the account types, in the Amazon Chime Administration Guide.

Previously suspended users who are unsuspended using this action are returned to Registered status. Users who are not previously suspended are ignored.

Arguments

account_id: The Amazon Chime account ID.

user_id_list: The request containing the user IDs to unsuspend.

Main.Chime.batch_update_phone_number — Function

batch_update_phone_number(update_phone_number_request_items)
batch_update_phone_number(update_phone_number_request_items, params::Dict{String,<:Any})

Updates phone number product types or calling names. You can update one attribute at a time for each UpdatePhoneNumberRequestItem. For example, you can update the product type or the calling name.

For toll-free numbers, you cannot use the Amazon Chime Business Calling product type. For numbers outside the U.S., you must use the Amazon Chime SIP Media Application Dial-In product type.

Updates to outbound calling names can take up to 72 hours to complete. Pending updates to outbound calling names must be complete before you can request another update.

Arguments

update_phone_number_request_items: The request containing the phone number IDs and product types or calling names to update.

Main.Chime.batch_update_user — Function

batch_update_user(account_id, update_user_request_items)
batch_update_user(account_id, update_user_request_items, params::Dict{String,<:Any})

Updates user details within the UpdateUserRequestItem object for up to 20 users for the specified Amazon Chime account. Currently, only LicenseType updates are supported for this action.

Arguments

account_id: The Amazon Chime account ID.

update_user_request_items: The request containing the user IDs and details to update.

Main.Chime.create_account — Function

create_account(name)
create_account(name, params::Dict{String,<:Any})

Creates an Amazon Chime account under the administrator's AWS account. Only Team account types are currently supported for this action. For more information about different account types, see Managing Your Amazon Chime Accounts in the Amazon Chime Administration Guide.

Arguments

name: The name of the Amazon Chime account.

Main.Chime.create_bot — Function

create_bot(account_id, display_name)
create_bot(account_id, display_name, params::Dict{String,<:Any})

Creates a bot for an Amazon Chime Enterprise account.

Arguments

account_id: The Amazon Chime account ID.

display_name: The bot display name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Domain": The domain of the Amazon Chime Enterprise account.

Main.Chime.create_meeting_dial_out — Function

create_meeting_dial_out(from_phone_number, join_token, meeting_id, to_phone_number)
create_meeting_dial_out(from_phone_number, join_token, meeting_id, to_phone_number, params::Dict{String,<:Any})

Uses the join token and call metadata in a meeting request (From number, To number, and so forth) to initiate an outbound call to a public switched telephone network (PSTN) and join them into a Chime meeting. Also ensures that the From number belongs to the customer.

To play welcome audio or implement an interactive voice response (IVR), use the CreateSipMediaApplicationCall action with the corresponding SIP media application ID.

Important

This API is not available in a dedicated namespace.

Arguments

from_phone_number: Phone number used as the caller ID when the remote party receives a call.

join_token: Token used by the Amazon Chime SDK attendee. Call the CreateAttendee action to get a join token.

meeting_id: The Amazon Chime SDK meeting ID.

to_phone_number: Phone number called when inviting someone to a meeting.

Main.Chime.create_phone_number_order — Function

create_phone_number_order(e164_phone_numbers, product_type)
create_phone_number_order(e164_phone_numbers, product_type, params::Dict{String,<:Any})

Creates an order for phone numbers to be provisioned. For toll-free numbers, you cannot use the Amazon Chime Business Calling product type. For numbers outside the U.S., you must use the Amazon Chime SIP Media Application Dial-In product type.

Arguments

e164_phone_numbers: List of phone numbers, in E.164 format.

product_type: The phone number product type.

Main.Chime.create_room — Function

create_room(account_id, name)
create_room(account_id, name, params::Dict{String,<:Any})

Creates a chat room for the specified Amazon Chime Enterprise account.

Arguments

account_id: The Amazon Chime account ID.

name: The room name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientRequestToken": The idempotency token for the request.

Main.Chime.create_room_membership — Function

create_room_membership(account_id, member_id, room_id)
create_room_membership(account_id, member_id, room_id, params::Dict{String,<:Any})

Adds a member to a chat room in an Amazon Chime Enterprise account. A member can be either a user or a bot. The member role designates whether the member is a chat room administrator or a general chat room member.

Arguments

account_id: The Amazon Chime account ID.

member_id: The Amazon Chime member ID (user ID or bot ID).

room_id: The room ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Role": The role of the member.

Main.Chime.create_user — Function

create_user(account_id)
create_user(account_id, params::Dict{String,<:Any})

Creates a user under the specified Amazon Chime account.

Arguments

account_id: The Amazon Chime account ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Email": The user's email address.

"UserType": The user type.

"Username": The user name.

Main.Chime.delete_account — Function

delete_account(account_id)
delete_account(account_id, params::Dict{String,<:Any})

Deletes the specified Amazon Chime account. You must suspend all users before deleting Team account. You can use the batch_suspend_user action to dodo.

For EnterpriseLWA and EnterpriseAD accounts, you must release the claimed domains for your Amazon Chime account before deletion. As soon as you release the domain, all users under that account are suspended.

Deleted accounts appear in your Disabled accounts list for 90 days. To restore deleted account from your Disabled accounts list, you must contact AWS Support.

After 90 days, deleted accounts are permanently removed from your Disabled accounts list.

Arguments

account_id: The Amazon Chime account ID.

Main.Chime.delete_events_configuration — Function

delete_events_configuration(account_id, bot_id)
delete_events_configuration(account_id, bot_id, params::Dict{String,<:Any})

Deletes the events configuration that allows a bot to receive outgoing events.

Arguments

account_id: The Amazon Chime account ID.

bot_id: The bot ID.

Main.Chime.delete_phone_number — Function

delete_phone_number(phone_number_id)
delete_phone_number(phone_number_id, params::Dict{String,<:Any})

Moves the specified phone number into the Deletion queue. A phone number must be disassociated from any users or Amazon Chime Voice Connectors before it can be deleted.

Deleted phone numbers remain in the Deletion queue for 7 days before they are deleted permanently.

Arguments

phone_number_id: The phone number ID.

Main.Chime.delete_room — Function

delete_room(account_id, room_id)
delete_room(account_id, room_id, params::Dict{String,<:Any})

Deletes a chat room in an Amazon Chime Enterprise account.

Arguments

account_id: The Amazon Chime account ID.

room_id: The chat room ID.

Main.Chime.delete_room_membership — Function

delete_room_membership(account_id, member_id, room_id)
delete_room_membership(account_id, member_id, room_id, params::Dict{String,<:Any})

Removes a member from a chat room in an Amazon Chime Enterprise account.

Arguments

account_id: The Amazon Chime account ID.

member_id: The member ID (user ID or bot ID).

room_id: The room ID.

Main.Chime.disassociate_phone_number_from_user — Function

disassociate_phone_number_from_user(account_id, user_id)
disassociate_phone_number_from_user(account_id, user_id, params::Dict{String,<:Any})

Disassociates the primary provisioned phone number from the specified Amazon Chime user.

Arguments

account_id: The Amazon Chime account ID.

user_id: The user ID.

Main.Chime.disassociate_signin_delegate_groups_from_account — Function

disassociate_signin_delegate_groups_from_account(account_id, group_names)
disassociate_signin_delegate_groups_from_account(account_id, group_names, params::Dict{String,<:Any})

Disassociates the specified sign-in delegate groups from the specified Amazon Chime account.

Arguments

account_id: The Amazon Chime account ID.

group_names: The sign-in delegate group names.

Main.Chime.get_account — Function

get_account(account_id)
get_account(account_id, params::Dict{String,<:Any})

Retrieves details for the specified Amazon Chime account, such as account type and supported licenses.

Arguments

account_id: The Amazon Chime account ID.

Main.Chime.get_account_settings — Function

get_account_settings(account_id)
get_account_settings(account_id, params::Dict{String,<:Any})

Retrieves account settings for the specified Amazon Chime account ID, such as remote control and dialout settings. For more information about these settings, see Use the Policies Page in the Amazon Chime Administration Guide.

Arguments

account_id: The Amazon Chime account ID.

Main.Chime.get_bot — Function

get_bot(account_id, bot_id)
get_bot(account_id, bot_id, params::Dict{String,<:Any})

Retrieves details for the specified bot, such as bot email address, bot type, status, and display name.

Arguments

account_id: The Amazon Chime account ID.

bot_id: The bot ID.

Main.Chime.get_events_configuration — Function

get_events_configuration(account_id, bot_id)
get_events_configuration(account_id, bot_id, params::Dict{String,<:Any})

Gets details for an events configuration that allows a bot to receive outgoing events, such as an HTTPS endpoint or Lambda function ARN.

Arguments

account_id: The Amazon Chime account ID.

bot_id: The bot ID.

Main.Chime.get_global_settings — Function

get_global_settings()
get_global_settings(params::Dict{String,<:Any})

Retrieves global settings for the administrator's AWS account, such as Amazon Chime Business Calling and Amazon Chime Voice Connector settings.

Main.Chime.get_phone_number — Function

get_phone_number(phone_number_id)
get_phone_number(phone_number_id, params::Dict{String,<:Any})

Retrieves details for the specified phone number ID, such as associations, capabilities, and product type.

Arguments

phone_number_id: The phone number ID.

Main.Chime.get_phone_number_order — Function

get_phone_number_order(phone_number_order_id)
get_phone_number_order(phone_number_order_id, params::Dict{String,<:Any})

Retrieves details for the specified phone number order, such as the order creation timestamp, phone numbers in E.164 format, product type, and order status.

Arguments

phone_number_order_id: The ID for the phone number order.

Main.Chime.get_phone_number_settings — Function

get_phone_number_settings()
get_phone_number_settings(params::Dict{String,<:Any})

Retrieves the phone number settings for the administrator's AWS account, such as the default outbound calling name.

Main.Chime.get_retention_settings — Function

get_retention_settings(account_id)
get_retention_settings(account_id, params::Dict{String,<:Any})

Gets the retention settings for the specified Amazon Chime Enterprise account. For more information about retention settings, see Managing Chat Retention Policies in the Amazon Chime Administration Guide.

Arguments

account_id: The Amazon Chime account ID.

Main.Chime.get_room — Function

get_room(account_id, room_id)
get_room(account_id, room_id, params::Dict{String,<:Any})

Retrieves room details, such as the room name, for a room in an Amazon Chime Enterprise account.

Arguments

account_id: The Amazon Chime account ID.

room_id: The room ID.

Main.Chime.get_user — Function

get_user(account_id, user_id)
get_user(account_id, user_id, params::Dict{String,<:Any})

Retrieves details for the specified user ID, such as primary email address, license type,and personal meeting PIN.

To retrieve user details with an email address instead of a user ID, use the list_users action, and then filter by email address.

Arguments

account_id: The Amazon Chime account ID.

user_id: The user ID.

Main.Chime.get_user_settings — Function

get_user_settings(account_id, user_id)
get_user_settings(account_id, user_id, params::Dict{String,<:Any})

Retrieves settings for the specified user ID, such as any associated phone number settings.

Arguments

account_id: The Amazon Chime account ID.

user_id: The user ID.

Main.Chime.invite_users — Function

invite_users(account_id, user_email_list)
invite_users(account_id, user_email_list, params::Dict{String,<:Any})

Sends email to a maximum of 50 users, inviting them to the specified Amazon Chime Team account. Only Team account types are currently supported for this action.

Arguments

account_id: The Amazon Chime account ID.

user_email_list: The user email addresses to which to send the email invitation.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"UserType": The user type.

Main.Chime.list_accounts — Function

list_accounts()
list_accounts(params::Dict{String,<:Any})

Lists the Amazon Chime accounts under the administrator's AWS account. You can filter accounts by account name prefix. To find out which Amazon Chime account a user belongs to, you can filter by the user's email address, which returns one account result.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"max-results": The maximum number of results to return in a single call. Defaults to

"name": Amazon Chime account name prefix with which to filter results.

"next-token": The token to use to retrieve the next page of results.

"user-email": User email address with which to filter results.

Main.Chime.list_bots — Function

list_bots(account_id)
list_bots(account_id, params::Dict{String,<:Any})

Lists the bots associated with the administrator's Amazon Chime Enterprise account ID.

Arguments

account_id: The Amazon Chime account ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"max-results": The maximum number of results to return in a single call. The default is

"next-token": The token to use to retrieve the next page of results.

Main.Chime.list_phone_number_orders — Function

list_phone_number_orders()
list_phone_number_orders(params::Dict{String,<:Any})

Lists the phone number orders for the administrator's Amazon Chime account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"max-results": The maximum number of results to return in a single call.

"next-token": The token to use to retrieve the next page of results.

Main.Chime.list_phone_numbers — Function

list_phone_numbers()
list_phone_numbers(params::Dict{String,<:Any})

Lists the phone numbers for the specified Amazon Chime account, Amazon Chime user, Amazon Chime Voice Connector, or Amazon Chime Voice Connector group.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filter-name": The filter to use to limit the number of results.

"filter-value": The value to use for the filter.

"max-results": The maximum number of results to return in a single call.

"next-token": The token to use to retrieve the next page of results.

"product-type": The phone number product type.

"status": The phone number status.

Main.Chime.list_room_memberships — Function

list_room_memberships(account_id, room_id)
list_room_memberships(account_id, room_id, params::Dict{String,<:Any})

Lists the membership details for the specified room in an Amazon Chime Enterprise account, such as the members' IDs, email addresses, and names.

Arguments

account_id: The Amazon Chime account ID.

room_id: The room ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"max-results": The maximum number of results to return in a single call.

"next-token": The token to use to retrieve the next page of results.

Main.Chime.list_rooms — Function

list_rooms(account_id)
list_rooms(account_id, params::Dict{String,<:Any})

Lists the room details for the specified Amazon Chime Enterprise account. Optionally, filter the results by a member ID (user ID or bot ID) to see a list of rooms that the member belongs to.

Arguments

account_id: The Amazon Chime account ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"max-results": The maximum number of results to return in a single call.

"member-id": The member ID (user ID or bot ID).

"next-token": The token to use to retrieve the next page of results.

Main.Chime.list_supported_phone_number_countries — Function

list_supported_phone_number_countries(product-type)
list_supported_phone_number_countries(product-type, params::Dict{String,<:Any})

Lists supported phone number countries.

Arguments

product-type: The phone number product type.

Main.Chime.list_users — Function

list_users(account_id)
list_users(account_id, params::Dict{String,<:Any})

Lists the users that belong to the specified Amazon Chime account. You can specify an email address to list only the user that the email address belongs to.

Arguments

account_id: The Amazon Chime account ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"max-results": The maximum number of results to return in a single call. Defaults to

"next-token": The token to use to retrieve the next page of results.

"user-email": Optional. The user email address used to filter results. Maximum 1.

"user-type": The user type.

Main.Chime.logout_user — Function

logout_user(account_id, user_id)
logout_user(account_id, user_id, params::Dict{String,<:Any})

Logs out the specified user from all of the devices they are currently logged into.

Arguments

account_id: The Amazon Chime account ID.

user_id: The user ID.

Main.Chime.put_events_configuration — Function

put_events_configuration(account_id, bot_id)
put_events_configuration(account_id, bot_id, params::Dict{String,<:Any})

Creates an events configuration that allows a bot to receive outgoing events sent by Amazon Chime. Choose either an HTTPS endpoint or a Lambda function ARN. For more information, see Bot.

Arguments

account_id: The Amazon Chime account ID.

bot_id: The bot ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"LambdaFunctionArn": Lambda function ARN that allows the bot to receive outgoing events.

"OutboundEventsHTTPSEndpoint": HTTPS endpoint that allows the bot to receive outgoing events.

Main.Chime.put_retention_settings — Function

put_retention_settings(account_id, retention_settings)
put_retention_settings(account_id, retention_settings, params::Dict{String,<:Any})

Puts retention settings for the specified Amazon Chime Enterprise account. We recommend using AWS CloudTrail to monitor usage of this API for your account. For more information, see Logging Amazon Chime API Calls with AWS CloudTrail in the Amazon Chime Administration Guide.

To turn off existing retention settings, remove the number of days from the corresponding RetentionDays field in the RetentionSettings object. For more information about retention settings, see Managing Chat Retention Policies in the Amazon Chime Administration Guide.

Arguments

account_id: The Amazon Chime account ID.

retention_settings: The retention settings.

Main.Chime.redact_conversation_message — Function

redact_conversation_message(account_id, conversation_id, message_id)
redact_conversation_message(account_id, conversation_id, message_id, params::Dict{String,<:Any})

Redacts the specified message from the specified Amazon Chime conversation.

Arguments

account_id: The Amazon Chime account ID.

conversation_id: The conversation ID.

message_id: The message ID.

Main.Chime.redact_room_message — Function

redact_room_message(account_id, message_id, room_id)
redact_room_message(account_id, message_id, room_id, params::Dict{String,<:Any})

Redacts the specified message from the specified Amazon Chime channel.

Arguments

account_id: The Amazon Chime account ID.

message_id: The message ID.

room_id: The room ID.

Main.Chime.regenerate_security_token — Function

regenerate_security_token(account_id, bot_id)
regenerate_security_token(account_id, bot_id, params::Dict{String,<:Any})

Regenerates the security token for a bot.

Arguments

account_id: The Amazon Chime account ID.

bot_id: The bot ID.

Main.Chime.reset_personal_pin — Function

reset_personal_pin(account_id, user_id)
reset_personal_pin(account_id, user_id, params::Dict{String,<:Any})

Resets the personal meeting PIN for the specified user on an Amazon Chime account. Returns the User object with the updated personal meeting PIN.

Arguments

account_id: The Amazon Chime account ID.

user_id: The user ID.

Main.Chime.restore_phone_number — Function

restore_phone_number(phone_number_id)
restore_phone_number(phone_number_id, params::Dict{String,<:Any})

Moves a phone number from the Deletion queue back into the phone number Inventory.

Arguments

phone_number_id: The phone number.

Main.Chime.search_available_phone_numbers — Function

search_available_phone_numbers()
search_available_phone_numbers(params::Dict{String,<:Any})

Searches for phone numbers that can be ordered. For US numbers, provide at least one of the following search filters: AreaCode, City, State, or TollFreePrefix. If you provide City, you must also provide State. Numbers outside the US only support the PhoneNumberType filter, which you must use.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"area-code": The area code used to filter results. Only applies to the US.

"city": The city used to filter results. Only applies to the US.

"country": The country used to filter results. Defaults to the US Format: ISO 3166-1 alpha-2.

"max-results": The maximum number of results to return in a single call.

"next-token": The token used to retrieve the next page of results.

"phone-number-type": The phone number type used to filter results. Required for non-US numbers.

"state": The state used to filter results. Required only if you provide City. Only applies to the US.

"toll-free-prefix": The toll-free prefix that you use to filter results. Only applies to the US.

Main.Chime.update_account — Function

update_account(account_id)
update_account(account_id, params::Dict{String,<:Any})

Updates account details for the specified Amazon Chime account. Currently, only account name and default license updates are supported for this action.

Arguments

account_id: The Amazon Chime account ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DefaultLicense": The default license applied when you add users to an Amazon Chime account.

"Name": The new name for the specified Amazon Chime account.

Main.Chime.update_account_settings — Function

update_account_settings(account_id, account_settings)
update_account_settings(account_id, account_settings, params::Dict{String,<:Any})

Updates the settings for the specified Amazon Chime account. You can update settings for remote control of shared screens, or for the dial-out option. For more information about these settings, see Use the Policies Page in the Amazon Chime Administration Guide.

Arguments

account_id: The Amazon Chime account ID.

account_settings: The Amazon Chime account settings to update.

Main.Chime.update_bot — Function

update_bot(account_id, bot_id)
update_bot(account_id, bot_id, params::Dict{String,<:Any})

Updates the status of the specified bot, such as starting or stopping the bot from running in your Amazon Chime Enterprise account.

Arguments

account_id: The Amazon Chime account ID.

bot_id: The bot ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Disabled": When true, stops the specified bot from running in your account.

Main.Chime.update_global_settings — Function

update_global_settings()
update_global_settings(params::Dict{String,<:Any})

Updates global settings for the administrator's AWS account, such as Amazon Chime Business Calling and Amazon Chime Voice Connector settings.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"BusinessCalling": The Amazon Chime Business Calling settings.

"VoiceConnector": The Amazon Chime Voice Connector settings.

Main.Chime.update_phone_number — Function

update_phone_number(phone_number_id)
update_phone_number(phone_number_id, params::Dict{String,<:Any})

Updates phone number details, such as product type or calling name, for the specified phone number ID. You can update one phone number detail at a time. For example, you can update either the product type or the calling name in one action.

For toll-free numbers, you cannot use the Amazon Chime Business Calling product type. For numbers outside the U.S., you must use the Amazon Chime SIP Media Application Dial-In product type.

Updates to outbound calling names can take 72 hours to complete. Pending updates to outbound calling names must be complete before you can request another update.

Arguments

phone_number_id: The phone number ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"CallingName": The outbound calling name associated with the phone number.

"ProductType": The product type.

Main.Chime.update_phone_number_settings — Function

update_phone_number_settings(calling_name)
update_phone_number_settings(calling_name, params::Dict{String,<:Any})

Updates the phone number settings for the administrator's AWS account, such as the default outbound calling name. You can update the default outbound calling name once every seven days. Outbound calling names can take up to 72 hours to update.

Arguments

calling_name: The default outbound calling name for the account.

Main.Chime.update_room — Function

update_room(account_id, room_id)
update_room(account_id, room_id, params::Dict{String,<:Any})

Updates room details, such as the room name, for a room in an Amazon Chime Enterprise account.

Arguments

account_id: The Amazon Chime account ID.

room_id: The room ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Name": The room name.

Main.Chime.update_room_membership — Function

update_room_membership(account_id, member_id, room_id)
update_room_membership(account_id, member_id, room_id, params::Dict{String,<:Any})

Updates room membership details, such as the member role, for a room in an Amazon Chime Enterprise account. The member role designates whether the member is a chat room administrator or a general chat room member. The member role can be updated only for user IDs.

Arguments

account_id: The Amazon Chime account ID.

member_id: The member ID.

room_id: The room ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Role": The role of the member.

Main.Chime.update_user — Function

update_user(account_id, user_id)
update_user(account_id, user_id, params::Dict{String,<:Any})

Updates user details for a specified user ID. Currently, only LicenseType updates are supported for this action.

Arguments

account_id: The Amazon Chime account ID.

user_id: The user ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AlexaForBusinessMetadata": The Alexa for Business metadata.

"LicenseType": The user license type to update. This must be a supported license type for the Amazon Chime account that the user belongs to.

"UserType": The user type.

Main.Chime.update_user_settings — Function

update_user_settings(account_id, user_id, user_settings)
update_user_settings(account_id, user_id, user_settings, params::Dict{String,<:Any})

Updates the settings for the specified user, such as phone number settings.

Arguments

account_id: The Amazon Chime account ID.

user_id: The user ID.

user_settings: The user settings to update.