Chime Sdk Identity

This page documents function available when using the Chime_Sdk_Identity module, created with @service Chime_Sdk_Identity.

Index

Main.Chime_Sdk_Identity.create_app_instance

Main.Chime_Sdk_Identity.create_app_instance_admin

Main.Chime_Sdk_Identity.create_app_instance_user

Main.Chime_Sdk_Identity.delete_app_instance

Main.Chime_Sdk_Identity.delete_app_instance_admin

Main.Chime_Sdk_Identity.delete_app_instance_user

Main.Chime_Sdk_Identity.deregister_app_instance_user_endpoint

Main.Chime_Sdk_Identity.describe_app_instance

Main.Chime_Sdk_Identity.describe_app_instance_admin

Main.Chime_Sdk_Identity.describe_app_instance_user

Main.Chime_Sdk_Identity.describe_app_instance_user_endpoint

Main.Chime_Sdk_Identity.get_app_instance_retention_settings

Main.Chime_Sdk_Identity.list_app_instance_admins

Main.Chime_Sdk_Identity.list_app_instance_user_endpoints

Main.Chime_Sdk_Identity.list_app_instance_users

Main.Chime_Sdk_Identity.list_app_instances

Main.Chime_Sdk_Identity.list_tags_for_resource

Main.Chime_Sdk_Identity.put_app_instance_retention_settings

Main.Chime_Sdk_Identity.register_app_instance_user_endpoint

Main.Chime_Sdk_Identity.tag_resource

Main.Chime_Sdk_Identity.untag_resource

Main.Chime_Sdk_Identity.update_app_instance

Main.Chime_Sdk_Identity.update_app_instance_user

Main.Chime_Sdk_Identity.update_app_instance_user_endpoint

Documentation

Main.Chime_Sdk_Identity.create_app_instance — Method

create_app_instance(client_request_token, name)
create_app_instance(client_request_token, name, params::Dict{String,<:Any})

Creates an Amazon Chime SDK messaging AppInstance under an AWS account. Only SDK messaging customers use this API. CreateAppInstance supports idempotency behavior as described in the AWS API Standard. identity

Arguments

client_request_token: The ClientRequestToken of the AppInstance.

name: The name of the AppInstance.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Metadata": The metadata of the AppInstance. Limited to a 1KB string in UTF-8.

"Tags": Tags assigned to the AppInstanceUser.

Main.Chime_Sdk_Identity.create_app_instance_admin — Method

create_app_instance_admin(app_instance_admin_arn, app_instance_arn)
create_app_instance_admin(app_instance_admin_arn, app_instance_arn, params::Dict{String,<:Any})

Promotes an AppInstanceUser to an AppInstanceAdmin. The promoted user can perform the following actions. ChannelModerator actions across all channels in the AppInstance. DeleteChannelMessage actions. Only an AppInstanceUser can be promoted to an AppInstanceAdmin role.

Arguments

app_instance_admin_arn: The ARN of the administrator of the current AppInstance.

app_instance_arn: The ARN of the AppInstance.

Main.Chime_Sdk_Identity.create_app_instance_user — Method

create_app_instance_user(app_instance_arn, app_instance_user_id, client_request_token, name)
create_app_instance_user(app_instance_arn, app_instance_user_id, client_request_token, name, params::Dict{String,<:Any})

Creates a user under an Amazon Chime AppInstance. The request consists of a unique appInstanceUserId and Name for that user.

Arguments

app_instance_arn: The ARN of the AppInstance request.

app_instance_user_id: The user ID of the AppInstance.

client_request_token: The token assigned to the user requesting an AppInstance.

name: The user's name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Metadata": The request's metadata. Limited to a 1KB string in UTF-8.

"Tags": Tags assigned to the AppInstanceUser.

Main.Chime_Sdk_Identity.delete_app_instance — Method

delete_app_instance(app_instance_arn)
delete_app_instance(app_instance_arn, params::Dict{String,<:Any})

Deletes an AppInstance and all associated data asynchronously.

Arguments

app_instance_arn: The ARN of the AppInstance.

Main.Chime_Sdk_Identity.delete_app_instance_admin — Method

delete_app_instance_admin(app_instance_admin_arn, app_instance_arn)
delete_app_instance_admin(app_instance_admin_arn, app_instance_arn, params::Dict{String,<:Any})

Demotes an AppInstanceAdmin to an AppInstanceUser. This action does not delete the user.

Arguments

app_instance_admin_arn: The ARN of the AppInstance's administrator.

app_instance_arn: The ARN of the AppInstance.

Main.Chime_Sdk_Identity.delete_app_instance_user — Method

delete_app_instance_user(app_instance_user_arn)
delete_app_instance_user(app_instance_user_arn, params::Dict{String,<:Any})

Deletes an AppInstanceUser.

Arguments

app_instance_user_arn: The ARN of the user request being deleted.

Main.Chime_Sdk_Identity.deregister_app_instance_user_endpoint — Method

deregister_app_instance_user_endpoint(app_instance_user_arn, endpoint_id)
deregister_app_instance_user_endpoint(app_instance_user_arn, endpoint_id, params::Dict{String,<:Any})

Deregisters an AppInstanceUserEndpoint.

Arguments

app_instance_user_arn: The ARN of the AppInstanceUser.

endpoint_id: The unique identifier of the AppInstanceUserEndpoint.

Main.Chime_Sdk_Identity.describe_app_instance — Method

describe_app_instance(app_instance_arn)
describe_app_instance(app_instance_arn, params::Dict{String,<:Any})

Returns the full details of an AppInstance.

Arguments

app_instance_arn: The ARN of the AppInstance.

Main.Chime_Sdk_Identity.describe_app_instance_admin — Method

describe_app_instance_admin(app_instance_admin_arn, app_instance_arn)
describe_app_instance_admin(app_instance_admin_arn, app_instance_arn, params::Dict{String,<:Any})

Returns the full details of an AppInstanceAdmin.

Arguments

app_instance_admin_arn: The ARN of the AppInstanceAdmin.

app_instance_arn: The ARN of the AppInstance.

Main.Chime_Sdk_Identity.describe_app_instance_user — Method

describe_app_instance_user(app_instance_user_arn)
describe_app_instance_user(app_instance_user_arn, params::Dict{String,<:Any})

Returns the full details of an AppInstanceUser.

Arguments

app_instance_user_arn: The ARN of the AppInstanceUser.

Main.Chime_Sdk_Identity.describe_app_instance_user_endpoint — Method

describe_app_instance_user_endpoint(app_instance_user_arn, endpoint_id)
describe_app_instance_user_endpoint(app_instance_user_arn, endpoint_id, params::Dict{String,<:Any})

Returns the full details of an AppInstanceUserEndpoint.

Arguments

app_instance_user_arn: The ARN of the AppInstanceUser.

endpoint_id: The unique identifier of the AppInstanceUserEndpoint.

Main.Chime_Sdk_Identity.get_app_instance_retention_settings — Method

get_app_instance_retention_settings(app_instance_arn)
get_app_instance_retention_settings(app_instance_arn, params::Dict{String,<:Any})

Gets the retention settings for an AppInstance.

Arguments

app_instance_arn: The ARN of the AppInstance.

Main.Chime_Sdk_Identity.list_app_instance_admins — Method

list_app_instance_admins(app_instance_arn)
list_app_instance_admins(app_instance_arn, params::Dict{String,<:Any})

Returns a list of the administrators in the AppInstance.

Arguments

app_instance_arn: The ARN of the AppInstance.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"max-results": The maximum number of administrators that you want to return.

"next-token": The token returned from previous API requests until the number of administrators is reached.

Main.Chime_Sdk_Identity.list_app_instance_user_endpoints — Method

list_app_instance_user_endpoints(app_instance_user_arn)
list_app_instance_user_endpoints(app_instance_user_arn, params::Dict{String,<:Any})

Lists all the AppInstanceUserEndpoints created under a single AppInstanceUser.

Arguments

app_instance_user_arn: The ARN of the AppInstanceUser.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"max-results": The maximum number of endpoints that you want to return.

"next-token": The token passed by previous API calls until all requested endpoints are returned.

Main.Chime_Sdk_Identity.list_app_instance_users — Method

list_app_instance_users(app-instance-arn)
list_app_instance_users(app-instance-arn, params::Dict{String,<:Any})

List all AppInstanceUsers created under a single AppInstance.

Arguments

app-instance-arn: The ARN of the AppInstance.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"max-results": The maximum number of requests that you want returned.

"next-token": The token passed by previous API calls until all requested users are returned.

Main.Chime_Sdk_Identity.list_app_instances — Method

list_app_instances()
list_app_instances(params::Dict{String,<:Any})

Lists all Amazon Chime AppInstances created under a single AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"max-results": The maximum number of AppInstances that you want to return.

"next-token": The token passed by previous API requests until you reach the maximum number of AppInstances.

Main.Chime_Sdk_Identity.list_tags_for_resource — Method

list_tags_for_resource(arn)
list_tags_for_resource(arn, params::Dict{String,<:Any})

Lists the tags applied to an Amazon Chime SDK identity resource.

Arguments

arn: The ARN of the resource.

Main.Chime_Sdk_Identity.put_app_instance_retention_settings — Method

put_app_instance_retention_settings(app_instance_retention_settings, app_instance_arn)
put_app_instance_retention_settings(app_instance_retention_settings, app_instance_arn, params::Dict{String,<:Any})

Sets the amount of time in days that a given AppInstance retains data.

Arguments

app_instance_retention_settings: The time in days to retain data. Data type: number.

app_instance_arn: The ARN of the AppInstance.

Main.Chime_Sdk_Identity.register_app_instance_user_endpoint — Method

register_app_instance_user_endpoint(client_request_token, endpoint_attributes, resource_arn, type, app_instance_user_arn)
register_app_instance_user_endpoint(client_request_token, endpoint_attributes, resource_arn, type, app_instance_user_arn, params::Dict{String,<:Any})

Registers an endpoint under an Amazon Chime AppInstanceUser. The endpoint receives messages for a user. For push notifications, the endpoint is a mobile device used to receive mobile push notifications for a user.

Arguments

client_request_token: The idempotency token for each client request.

endpoint_attributes: The attributes of an Endpoint.

resource_arn: The ARN of the resource to which the endpoint belongs.

type: The type of the AppInstanceUserEndpoint. Supported types: APNS: The mobile notification service for an Apple device. APNS_SANDBOX: The sandbox environment of the mobile notification service for an Apple device. GCM: The mobile notification service for an Android device. Populate the ResourceArn value of each type as PinpointAppArn.

app_instance_user_arn: The ARN of the AppInstanceUser.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AllowMessages": Boolean that controls whether the AppInstanceUserEndpoint is opted in to receive messages. ALL indicates the endpoint receives all messages. NONE indicates the endpoint receives no messages.

"Name": The name of the AppInstanceUserEndpoint.

Main.Chime_Sdk_Identity.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Applies the specified tags to the specified Amazon Chime SDK identity resource.

Arguments

resource_arn: The resource ARN.

tags: The tag key-value pairs.

Main.Chime_Sdk_Identity.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes the specified tags from the specified Amazon Chime SDK identity resource.

Arguments

resource_arn: The resource ARN.

tag_keys: The tag keys.

Main.Chime_Sdk_Identity.update_app_instance — Method

update_app_instance(metadata, name, app_instance_arn)
update_app_instance(metadata, name, app_instance_arn, params::Dict{String,<:Any})

Updates AppInstance metadata.

Arguments

metadata: The metadata that you want to change.

name: The name that you want to change.

app_instance_arn: The ARN of the AppInstance.

Main.Chime_Sdk_Identity.update_app_instance_user — Method

update_app_instance_user(metadata, name, app_instance_user_arn)
update_app_instance_user(metadata, name, app_instance_user_arn, params::Dict{String,<:Any})

Updates the details of an AppInstanceUser. You can update names and metadata.

Arguments

metadata: The metadata of the AppInstanceUser.

name: The name of the AppInstanceUser.

app_instance_user_arn: The ARN of the AppInstanceUser.

Main.Chime_Sdk_Identity.update_app_instance_user_endpoint — Method

update_app_instance_user_endpoint(app_instance_user_arn, endpoint_id)
update_app_instance_user_endpoint(app_instance_user_arn, endpoint_id, params::Dict{String,<:Any})

Updates the details of an AppInstanceUserEndpoint. You can update the name and AllowMessage values.

Arguments

app_instance_user_arn: The ARN of the AppInstanceUser.

endpoint_id: The unique identifier of the AppInstanceUserEndpoint.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AllowMessages": Boolean that controls whether the AppInstanceUserEndpoint is opted in to receive messages. ALL indicates the endpoint will receive all messages. NONE indicates the endpoint will receive no messages.

"Name": The name of the AppInstanceUserEndpoint.