Cloudfront

This page documents function available when using the Cloudfront module, created with @service Cloudfront.

Index

Main.Cloudfront.create_cache_policy2020_05_31

Main.Cloudfront.create_cloud_front_origin_access_identity2020_05_31

Main.Cloudfront.create_distribution2020_05_31

Main.Cloudfront.create_distribution_with_tags2020_05_31

Main.Cloudfront.create_field_level_encryption_config2020_05_31

Main.Cloudfront.create_field_level_encryption_profile2020_05_31

Main.Cloudfront.create_invalidation2020_05_31

Main.Cloudfront.create_key_group2020_05_31

Main.Cloudfront.create_monitoring_subscription2020_05_31

Main.Cloudfront.create_origin_request_policy2020_05_31

Main.Cloudfront.create_public_key2020_05_31

Main.Cloudfront.create_realtime_log_config2020_05_31

Main.Cloudfront.create_streaming_distribution2020_05_31

Main.Cloudfront.create_streaming_distribution_with_tags2020_05_31

Main.Cloudfront.delete_cache_policy2020_05_31

Main.Cloudfront.delete_cloud_front_origin_access_identity2020_05_31

Main.Cloudfront.delete_distribution2020_05_31

Main.Cloudfront.delete_field_level_encryption_config2020_05_31

Main.Cloudfront.delete_field_level_encryption_profile2020_05_31

Main.Cloudfront.delete_key_group2020_05_31

Main.Cloudfront.delete_monitoring_subscription2020_05_31

Main.Cloudfront.delete_origin_request_policy2020_05_31

Main.Cloudfront.delete_public_key2020_05_31

Main.Cloudfront.delete_realtime_log_config2020_05_31

Main.Cloudfront.delete_streaming_distribution2020_05_31

Main.Cloudfront.get_cache_policy2020_05_31

Main.Cloudfront.get_cache_policy_config2020_05_31

Main.Cloudfront.get_cloud_front_origin_access_identity2020_05_31

Main.Cloudfront.get_cloud_front_origin_access_identity_config2020_05_31

Main.Cloudfront.get_distribution2020_05_31

Main.Cloudfront.get_distribution_config2020_05_31

Main.Cloudfront.get_field_level_encryption2020_05_31

Main.Cloudfront.get_field_level_encryption_config2020_05_31

Main.Cloudfront.get_field_level_encryption_profile2020_05_31

Main.Cloudfront.get_field_level_encryption_profile_config2020_05_31

Main.Cloudfront.get_invalidation2020_05_31

Main.Cloudfront.get_key_group2020_05_31

Main.Cloudfront.get_key_group_config2020_05_31

Main.Cloudfront.get_monitoring_subscription2020_05_31

Main.Cloudfront.get_origin_request_policy2020_05_31

Main.Cloudfront.get_origin_request_policy_config2020_05_31

Main.Cloudfront.get_public_key2020_05_31

Main.Cloudfront.get_public_key_config2020_05_31

Main.Cloudfront.get_realtime_log_config2020_05_31

Main.Cloudfront.get_streaming_distribution2020_05_31

Main.Cloudfront.get_streaming_distribution_config2020_05_31

Main.Cloudfront.list_cache_policies2020_05_31

Main.Cloudfront.list_cloud_front_origin_access_identities2020_05_31

Main.Cloudfront.list_distributions2020_05_31

Main.Cloudfront.list_distributions_by_cache_policy_id2020_05_31

Main.Cloudfront.list_distributions_by_key_group2020_05_31

Main.Cloudfront.list_distributions_by_origin_request_policy_id2020_05_31

Main.Cloudfront.list_distributions_by_realtime_log_config2020_05_31

Main.Cloudfront.list_distributions_by_web_aclid2020_05_31

Main.Cloudfront.list_field_level_encryption_configs2020_05_31

Main.Cloudfront.list_field_level_encryption_profiles2020_05_31

Main.Cloudfront.list_invalidations2020_05_31

Main.Cloudfront.list_key_groups2020_05_31

Main.Cloudfront.list_origin_request_policies2020_05_31

Main.Cloudfront.list_public_keys2020_05_31

Main.Cloudfront.list_realtime_log_configs2020_05_31

Main.Cloudfront.list_streaming_distributions2020_05_31

Main.Cloudfront.list_tags_for_resource2020_05_31

Main.Cloudfront.tag_resource2020_05_31

Main.Cloudfront.untag_resource2020_05_31

Main.Cloudfront.update_cache_policy2020_05_31

Main.Cloudfront.update_cloud_front_origin_access_identity2020_05_31

Main.Cloudfront.update_distribution2020_05_31

Main.Cloudfront.update_field_level_encryption_config2020_05_31

Main.Cloudfront.update_field_level_encryption_profile2020_05_31

Main.Cloudfront.update_key_group2020_05_31

Main.Cloudfront.update_origin_request_policy2020_05_31

Main.Cloudfront.update_public_key2020_05_31

Main.Cloudfront.update_realtime_log_config2020_05_31

Main.Cloudfront.update_streaming_distribution2020_05_31

Documentation

Main.Cloudfront.create_cache_policy2020_05_31 — Method

CreateCachePolicy2020_05_31()

Creates a cache policy. After you create a cache policy, you can attach it to one or more cache behaviors. When it’s attached to a cache behavior, the cache policy determines the following: The values that CloudFront includes in the cache key. These values can include HTTP headers, cookies, and URL query strings. CloudFront uses the cache key to find an object in its cache that it can return to the viewer. The default, minimum, and maximum time to live (TTL) values that you want objects to stay in the CloudFront cache. The headers, cookies, and query strings that are included in the cache key are automatically included in requests that CloudFront sends to the origin. CloudFront sends a request when it can’t find an object in its cache that matches the request’s cache key. If you want to send values to the origin but not include them in the cache key, use OriginRequestPolicy. For more information about cache policies, see Controlling the cache key in the Amazon CloudFront Developer Guide.

Required Parameters

CachePolicyConfig: A cache policy configuration.

Main.Cloudfront.create_cloud_front_origin_access_identity2020_05_31 — Method

CreateCloudFrontOriginAccessIdentity2020_05_31()

Creates a new origin access identity. If you're using Amazon S3 for your origin, you can use an origin access identity to require users to access your content using a CloudFront URL instead of the Amazon S3 URL. For more information about how to use origin access identities, see Serving Private Content through CloudFront in the Amazon CloudFront Developer Guide.

Required Parameters

CloudFrontOriginAccessIdentityConfig: The current configuration information for the identity.

Main.Cloudfront.create_distribution2020_05_31 — Method

CreateDistribution2020_05_31()

Creates a new web distribution. You create a CloudFront distribution to tell CloudFront where you want content to be delivered from, and the details about how to track and manage content delivery. Send a POST request to the /CloudFront API version/distribution/distribution ID resource. When you update a distribution, there are more required fields than when you create a distribution. When you update your distribution by using UpdateDistribution, follow the steps included in the documentation to get the current configuration and then make your updates. This helps to make sure that you include all of the required fields. To view a summary, see Required Fields for Create Distribution and Update Distribution in the Amazon CloudFront Developer Guide.

Required Parameters

DistributionConfig: The distribution's configuration information.

Main.Cloudfront.create_distribution_with_tags2020_05_31 — Method

CreateDistributionWithTags2020_05_31()

Create a new distribution with tags.

Required Parameters

DistributionConfigWithTags: The distribution's configuration information.

Main.Cloudfront.create_field_level_encryption_config2020_05_31 — Method

CreateFieldLevelEncryptionConfig2020_05_31()

Create a new field-level encryption configuration.

Required Parameters

FieldLevelEncryptionConfig: The request to create a new field-level encryption configuration.

Main.Cloudfront.create_field_level_encryption_profile2020_05_31 — Method

CreateFieldLevelEncryptionProfile2020_05_31()

Create a field-level encryption profile.

Required Parameters

FieldLevelEncryptionProfileConfig: The request to create a field-level encryption profile.

Main.Cloudfront.create_invalidation2020_05_31 — Method

CreateInvalidation2020_05_31()

Create a new invalidation.

Required Parameters

DistributionId: The distribution's id.

InvalidationBatch: The batch information for the invalidation.

Main.Cloudfront.create_key_group2020_05_31 — Method

CreateKeyGroup2020_05_31()

Creates a key group that you can use with CloudFront signed URLs and signed cookies. To create a key group, you must specify at least one public key for the key group. After you create a key group, you can reference it from one or more cache behaviors. When you reference a key group in a cache behavior, CloudFront requires signed URLs or signed cookies for all requests that match the cache behavior. The URLs or cookies must be signed with a private key whose corresponding public key is in the key group. The signed URL or cookie contains information about which public key CloudFront should use to verify the signature. For more information, see Serving private content in the Amazon CloudFront Developer Guide.

Required Parameters

KeyGroupConfig: A key group configuration.

Main.Cloudfront.create_monitoring_subscription2020_05_31 — Method

CreateMonitoringSubscription2020_05_31()

Enables additional CloudWatch metrics for the specified CloudFront distribution. The additional metrics incur an additional cost. For more information, see Viewing additional CloudFront distribution metrics in the Amazon CloudFront Developer Guide.

Required Parameters

DistributionId: The ID of the distribution that you are enabling metrics for.

MonitoringSubscription: A monitoring subscription. This structure contains information about whether additional CloudWatch metrics are enabled for a given CloudFront distribution.

Main.Cloudfront.create_origin_request_policy2020_05_31 — Method

CreateOriginRequestPolicy2020_05_31()

Creates an origin request policy. After you create an origin request policy, you can attach it to one or more cache behaviors. When it’s attached to a cache behavior, the origin request policy determines the values that CloudFront includes in requests that it sends to the origin. Each request that CloudFront sends to the origin includes the following: The request body and the URL path (without the domain name) from the viewer request. The headers that CloudFront automatically includes in every origin request, including Host, User-Agent, and X-Amz-Cf-Id. All HTTP headers, cookies, and URL query strings that are specified in the cache policy or the origin request policy. These can include items from the viewer request and, in the case of headers, additional ones that are added by CloudFront. CloudFront sends a request when it can’t find a valid object in its cache that matches the request. If you want to send values to the origin and also include them in the cache key, use CachePolicy. For more information about origin request policies, see Controlling origin requests in the Amazon CloudFront Developer Guide.

Required Parameters

OriginRequestPolicyConfig: An origin request policy configuration.

Main.Cloudfront.create_public_key2020_05_31 — Method

CreatePublicKey2020_05_31()

Uploads a public key to CloudFront that you can use with signed URLs and signed cookies, or with field-level encryption.

Required Parameters

PublicKeyConfig: A CloudFront public key configuration.

Main.Cloudfront.create_realtime_log_config2020_05_31 — Method

CreateRealtimeLogConfig2020_05_31()

Creates a real-time log configuration. After you create a real-time log configuration, you can attach it to one or more cache behaviors to send real-time log data to the specified Amazon Kinesis data stream. For more information about real-time log configurations, see Real-time logs in the Amazon CloudFront Developer Guide.

Required Parameters

EndPoints: Contains information about the Amazon Kinesis data stream where you are sending real-time log data.

Field: A list of fields to include in each real-time log record. For more information about fields, see Real-time log configuration fields in the Amazon CloudFront Developer Guide.

Name: A unique name to identify this real-time log configuration.

SamplingRate: The sampling rate for this real-time log configuration. The sampling rate determines the percentage of viewer requests that are represented in the real-time log data. You must provide an integer between 1 and 100, inclusive.

Optional Parameters

Fields: A list of fields to include in each real-time log record. For more information about fields, see Real-time log configuration fields in the Amazon CloudFront Developer Guide.

Main.Cloudfront.create_streaming_distribution2020_05_31 — Method

CreateStreamingDistribution2020_05_31()

This API is deprecated. Amazon CloudFront is deprecating real-time messaging protocol (RTMP) distributions on December 31, 2020. For more information, read the announcement on the Amazon CloudFront discussion forum.

Required Parameters

StreamingDistributionConfig: The streaming distribution's configuration information.

Main.Cloudfront.create_streaming_distribution_with_tags2020_05_31 — Method

CreateStreamingDistributionWithTags2020_05_31()

This API is deprecated. Amazon CloudFront is deprecating real-time messaging protocol (RTMP) distributions on December 31, 2020. For more information, read the announcement on the Amazon CloudFront discussion forum.

Required Parameters

StreamingDistributionConfigWithTags: The streaming distribution's configuration information.

Main.Cloudfront.delete_cache_policy2020_05_31 — Method

DeleteCachePolicy2020_05_31()

Deletes a cache policy. You cannot delete a cache policy if it’s attached to a cache behavior. First update your distributions to remove the cache policy from all cache behaviors, then delete the cache policy. To delete a cache policy, you must provide the policy’s identifier and version. To get these values, you can use ListCachePolicies or GetCachePolicy.

Required Parameters

Id: The unique identifier for the cache policy that you are deleting. To get the identifier, you can use ListCachePolicies.

Optional Parameters

If-Match: The version of the cache policy that you are deleting. The version is the cache policy’s ETag value, which you can get using ListCachePolicies, GetCachePolicy, or GetCachePolicyConfig.

Main.Cloudfront.delete_cloud_front_origin_access_identity2020_05_31 — Method

DeleteCloudFrontOriginAccessIdentity2020_05_31()

Delete an origin access identity.

Required Parameters

Id: The origin access identity's ID.

Optional Parameters

If-Match: The value of the ETag header you received from a previous GET or PUT request. For example: E2QWRUHAPOMQZL.

Main.Cloudfront.delete_distribution2020_05_31 — Method

DeleteDistribution2020_05_31()

Delete a distribution.

Required Parameters

Id: The distribution ID.

Optional Parameters

If-Match: The value of the ETag header that you received when you disabled the distribution. For example: E2QWRUHAPOMQZL.

Main.Cloudfront.delete_field_level_encryption_config2020_05_31 — Method

DeleteFieldLevelEncryptionConfig2020_05_31()

Remove a field-level encryption configuration.

Required Parameters

Id: The ID of the configuration you want to delete from CloudFront.

Optional Parameters

If-Match: The value of the ETag header that you received when retrieving the configuration identity to delete. For example: E2QWRUHAPOMQZL.

Main.Cloudfront.delete_field_level_encryption_profile2020_05_31 — Method

DeleteFieldLevelEncryptionProfile2020_05_31()

Remove a field-level encryption profile.

Required Parameters

Id: Request the ID of the profile you want to delete from CloudFront.

Optional Parameters

If-Match: The value of the ETag header that you received when retrieving the profile to delete. For example: E2QWRUHAPOMQZL.

Main.Cloudfront.delete_key_group2020_05_31 — Method

DeleteKeyGroup2020_05_31()

Deletes a key group. You cannot delete a key group that is referenced in a cache behavior. First update your distributions to remove the key group from all cache behaviors, then delete the key group. To delete a key group, you must provide the key group’s identifier and version. To get these values, use ListKeyGroups followed by GetKeyGroup or GetKeyGroupConfig.

Required Parameters

Id: The identifier of the key group that you are deleting. To get the identifier, use ListKeyGroups.

Optional Parameters

If-Match: The version of the key group that you are deleting. The version is the key group’s ETag value. To get the ETag, use GetKeyGroup or GetKeyGroupConfig.

Main.Cloudfront.delete_monitoring_subscription2020_05_31 — Method

DeleteMonitoringSubscription2020_05_31()

Disables additional CloudWatch metrics for the specified CloudFront distribution.

Required Parameters

DistributionId: The ID of the distribution that you are disabling metrics for.

Main.Cloudfront.delete_origin_request_policy2020_05_31 — Method

DeleteOriginRequestPolicy2020_05_31()

Deletes an origin request policy. You cannot delete an origin request policy if it’s attached to any cache behaviors. First update your distributions to remove the origin request policy from all cache behaviors, then delete the origin request policy. To delete an origin request policy, you must provide the policy’s identifier and version. To get the identifier, you can use ListOriginRequestPolicies or GetOriginRequestPolicy.

Required Parameters

Id: The unique identifier for the origin request policy that you are deleting. To get the identifier, you can use ListOriginRequestPolicies.

Optional Parameters

If-Match: The version of the origin request policy that you are deleting. The version is the origin request policy’s ETag value, which you can get using ListOriginRequestPolicies, GetOriginRequestPolicy, or GetOriginRequestPolicyConfig.

Main.Cloudfront.delete_public_key2020_05_31 — Method

DeletePublicKey2020_05_31()

Remove a public key you previously added to CloudFront.

Required Parameters

Id: The ID of the public key you want to remove from CloudFront.

Optional Parameters

If-Match: The value of the ETag header that you received when retrieving the public key identity to delete. For example: E2QWRUHAPOMQZL.

Main.Cloudfront.delete_realtime_log_config2020_05_31 — Method

DeleteRealtimeLogConfig2020_05_31()

Deletes a real-time log configuration. You cannot delete a real-time log configuration if it’s attached to a cache behavior. First update your distributions to remove the real-time log configuration from all cache behaviors, then delete the real-time log configuration. To delete a real-time log configuration, you can provide the configuration’s name or its Amazon Resource Name (ARN). You must provide at least one. If you provide both, CloudFront uses the name to identify the real-time log configuration to delete.

Optional Parameters

ARN: The Amazon Resource Name (ARN) of the real-time log configuration to delete.

Name: The name of the real-time log configuration to delete.

Main.Cloudfront.delete_streaming_distribution2020_05_31 — Method

DeleteStreamingDistribution2020_05_31()

Delete a streaming distribution. To delete an RTMP distribution using the CloudFront API, perform the following steps. To delete an RTMP distribution using the CloudFront API: Disable the RTMP distribution. Submit a GET Streaming Distribution Config request to get the current configuration and the Etag header for the distribution. Update the XML document that was returned in the response to your GET Streaming Distribution Config request to change the value of Enabled to false. Submit a PUT Streaming Distribution Config request to update the configuration for your distribution. In the request body, include the XML document that you updated in Step 3. Then set the value of the HTTP If-Match header to the value of the ETag header that CloudFront returned when you submitted the GET Streaming Distribution Config request in Step 2. Review the response to the PUT Streaming Distribution Config request to confirm that the distribution was successfully disabled. Submit a GET Streaming Distribution Config request to confirm that your changes have propagated. When propagation is complete, the value of Status is Deployed. Submit a DELETE Streaming Distribution request. Set the value of the HTTP If-Match header to the value of the ETag header that CloudFront returned when you submitted the GET Streaming Distribution Config request in Step 2. Review the response to your DELETE Streaming Distribution request to confirm that the distribution was successfully deleted. For information about deleting a distribution using the CloudFront console, see Deleting a Distribution in the Amazon CloudFront Developer Guide.

Required Parameters

Id: The distribution ID.

Optional Parameters

If-Match: The value of the ETag header that you received when you disabled the streaming distribution. For example: E2QWRUHAPOMQZL.

Main.Cloudfront.get_cache_policy2020_05_31 — Method

GetCachePolicy2020_05_31()

Gets a cache policy, including the following metadata: The policy’s identifier. The date and time when the policy was last modified. To get a cache policy, you must provide the policy’s identifier. If the cache policy is attached to a distribution’s cache behavior, you can get the policy’s identifier using ListDistributions or GetDistribution. If the cache policy is not attached to a cache behavior, you can get the identifier using ListCachePolicies.

Required Parameters

Id: The unique identifier for the cache policy. If the cache policy is attached to a distribution’s cache behavior, you can get the policy’s identifier using ListDistributions or GetDistribution. If the cache policy is not attached to a cache behavior, you can get the identifier using ListCachePolicies.

Main.Cloudfront.get_cache_policy_config2020_05_31 — Method

GetCachePolicyConfig2020_05_31()

Gets a cache policy configuration. To get a cache policy configuration, you must provide the policy’s identifier. If the cache policy is attached to a distribution’s cache behavior, you can get the policy’s identifier using ListDistributions or GetDistribution. If the cache policy is not attached to a cache behavior, you can get the identifier using ListCachePolicies.

Required Parameters

Id: The unique identifier for the cache policy. If the cache policy is attached to a distribution’s cache behavior, you can get the policy’s identifier using ListDistributions or GetDistribution. If the cache policy is not attached to a cache behavior, you can get the identifier using ListCachePolicies.

Main.Cloudfront.get_cloud_front_origin_access_identity2020_05_31 — Method

GetCloudFrontOriginAccessIdentity2020_05_31()

Get the information about an origin access identity.

Required Parameters

Id: The identity's ID.

Main.Cloudfront.get_cloud_front_origin_access_identity_config2020_05_31 — Method

GetCloudFrontOriginAccessIdentityConfig2020_05_31()

Get the configuration information about an origin access identity.

Required Parameters

Id: The identity's ID.

Main.Cloudfront.get_distribution2020_05_31 — Method

GetDistribution2020_05_31()

Get the information about a distribution.

Required Parameters

Id: The distribution's ID. If the ID is empty, an empty distribution configuration is returned.

Main.Cloudfront.get_distribution_config2020_05_31 — Method

GetDistributionConfig2020_05_31()

Get the configuration information about a distribution.

Required Parameters

Id: The distribution's ID. If the ID is empty, an empty distribution configuration is returned.

Main.Cloudfront.get_field_level_encryption2020_05_31 — Method

GetFieldLevelEncryption2020_05_31()

Get the field-level encryption configuration information.

Required Parameters

Id: Request the ID for the field-level encryption configuration information.

Main.Cloudfront.get_field_level_encryption_config2020_05_31 — Method

GetFieldLevelEncryptionConfig2020_05_31()

Get the field-level encryption configuration information.

Required Parameters

Id: Request the ID for the field-level encryption configuration information.

Main.Cloudfront.get_field_level_encryption_profile2020_05_31 — Method

GetFieldLevelEncryptionProfile2020_05_31()

Get the field-level encryption profile information.

Required Parameters

Id: Get the ID for the field-level encryption profile information.

Main.Cloudfront.get_field_level_encryption_profile_config2020_05_31 — Method

GetFieldLevelEncryptionProfileConfig2020_05_31()

Get the field-level encryption profile configuration information.

Required Parameters

Id: Get the ID for the field-level encryption profile configuration information.

Main.Cloudfront.get_invalidation2020_05_31 — Method

GetInvalidation2020_05_31()

Get the information about an invalidation.

Required Parameters

DistributionId: The distribution's ID.

Id: The identifier for the invalidation request, for example, IDFDVBD632BHDS5.

Main.Cloudfront.get_key_group2020_05_31 — Method

GetKeyGroup2020_05_31()

Gets a key group, including the date and time when the key group was last modified. To get a key group, you must provide the key group’s identifier. If the key group is referenced in a distribution’s cache behavior, you can get the key group’s identifier using ListDistributions or GetDistribution. If the key group is not referenced in a cache behavior, you can get the identifier using ListKeyGroups.

Required Parameters

Id: The identifier of the key group that you are getting. To get the identifier, use ListKeyGroups.

Main.Cloudfront.get_key_group_config2020_05_31 — Method

GetKeyGroupConfig2020_05_31()

Gets a key group configuration. To get a key group configuration, you must provide the key group’s identifier. If the key group is referenced in a distribution’s cache behavior, you can get the key group’s identifier using ListDistributions or GetDistribution. If the key group is not referenced in a cache behavior, you can get the identifier using ListKeyGroups.

Required Parameters

Id: The identifier of the key group whose configuration you are getting. To get the identifier, use ListKeyGroups.

Main.Cloudfront.get_monitoring_subscription2020_05_31 — Method

GetMonitoringSubscription2020_05_31()

Gets information about whether additional CloudWatch metrics are enabled for the specified CloudFront distribution.

Required Parameters

DistributionId: The ID of the distribution that you are getting metrics information for.

Main.Cloudfront.get_origin_request_policy2020_05_31 — Method

GetOriginRequestPolicy2020_05_31()

Gets an origin request policy, including the following metadata: The policy’s identifier. The date and time when the policy was last modified. To get an origin request policy, you must provide the policy’s identifier. If the origin request policy is attached to a distribution’s cache behavior, you can get the policy’s identifier using ListDistributions or GetDistribution. If the origin request policy is not attached to a cache behavior, you can get the identifier using ListOriginRequestPolicies.

Required Parameters

Id: The unique identifier for the origin request policy. If the origin request policy is attached to a distribution’s cache behavior, you can get the policy’s identifier using ListDistributions or GetDistribution. If the origin request policy is not attached to a cache behavior, you can get the identifier using ListOriginRequestPolicies.

Main.Cloudfront.get_origin_request_policy_config2020_05_31 — Method

GetOriginRequestPolicyConfig2020_05_31()

Gets an origin request policy configuration. To get an origin request policy configuration, you must provide the policy’s identifier. If the origin request policy is attached to a distribution’s cache behavior, you can get the policy’s identifier using ListDistributions or GetDistribution. If the origin request policy is not attached to a cache behavior, you can get the identifier using ListOriginRequestPolicies.

Required Parameters

Id: The unique identifier for the origin request policy. If the origin request policy is attached to a distribution’s cache behavior, you can get the policy’s identifier using ListDistributions or GetDistribution. If the origin request policy is not attached to a cache behavior, you can get the identifier using ListOriginRequestPolicies.

Main.Cloudfront.get_public_key2020_05_31 — Method

GetPublicKey2020_05_31()

Gets a public key.

Required Parameters

Id: The identifier of the public key you are getting.

Main.Cloudfront.get_public_key_config2020_05_31 — Method

GetPublicKeyConfig2020_05_31()

Gets a public key configuration.

Required Parameters

Id: The identifier of the public key whose configuration you are getting.

Main.Cloudfront.get_realtime_log_config2020_05_31 — Method

GetRealtimeLogConfig2020_05_31()

Gets a real-time log configuration. To get a real-time log configuration, you can provide the configuration’s name or its Amazon Resource Name (ARN). You must provide at least one. If you provide both, CloudFront uses the name to identify the real-time log configuration to get.

Optional Parameters

ARN: The Amazon Resource Name (ARN) of the real-time log configuration to get.

Name: The name of the real-time log configuration to get.

Main.Cloudfront.get_streaming_distribution2020_05_31 — Method

GetStreamingDistribution2020_05_31()

Gets information about a specified RTMP distribution, including the distribution configuration.

Required Parameters

Id: The streaming distribution's ID.

Main.Cloudfront.get_streaming_distribution_config2020_05_31 — Method

GetStreamingDistributionConfig2020_05_31()

Get the configuration information about a streaming distribution.

Required Parameters

Id: The streaming distribution's ID.

Main.Cloudfront.list_cache_policies2020_05_31 — Method

ListCachePolicies2020_05_31()

Gets a list of cache policies. You can optionally apply a filter to return only the managed policies created by AWS, or only the custom policies created in your AWS account. You can optionally specify the maximum number of items to receive in the response. If the total number of items in the list exceeds the maximum that you specify, or the default maximum, the response is paginated. To get the next page of items, send a subsequent request that specifies the NextMarker value from the current response as the Marker value in the subsequent request.

Optional Parameters

Marker: Use this field when paginating results to indicate where to begin in your list of cache policies. The response includes cache policies in the list that occur after the marker. To get the next page of the list, set this field’s value to the value of NextMarker from the current page’s response.

MaxItems: The maximum number of cache policies that you want in the response.

Type: A filter to return only the specified kinds of cache policies. Valid values are: managed – Returns only the managed policies created by AWS. custom – Returns only the custom policies created in your AWS account.

Main.Cloudfront.list_cloud_front_origin_access_identities2020_05_31 — Method

ListCloudFrontOriginAccessIdentities2020_05_31()

Lists origin access identities.

Optional Parameters

Marker: Use this when paginating results to indicate where to begin in your list of origin access identities. The results include identities in the list that occur after the marker. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response (which is also the ID of the last identity on that page).

MaxItems: The maximum number of origin access identities you want in the response body.

Main.Cloudfront.list_distributions2020_05_31 — Method

ListDistributions2020_05_31()

List CloudFront distributions.

Optional Parameters

Marker: Use this when paginating results to indicate where to begin in your list of distributions. The results include distributions in the list that occur after the marker. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response (which is also the ID of the last distribution on that page).

MaxItems: The maximum number of distributions you want in the response body.

Main.Cloudfront.list_distributions_by_cache_policy_id2020_05_31 — Method

ListDistributionsByCachePolicyId2020_05_31()

Gets a list of distribution IDs for distributions that have a cache behavior that’s associated with the specified cache policy. You can optionally specify the maximum number of items to receive in the response. If the total number of items in the list exceeds the maximum that you specify, or the default maximum, the response is paginated. To get the next page of items, send a subsequent request that specifies the NextMarker value from the current response as the Marker value in the subsequent request.

Required Parameters

CachePolicyId: The ID of the cache policy whose associated distribution IDs you want to list.

Optional Parameters

Marker: Use this field when paginating results to indicate where to begin in your list of distribution IDs. The response includes distribution IDs in the list that occur after the marker. To get the next page of the list, set this field’s value to the value of NextMarker from the current page’s response.

MaxItems: The maximum number of distribution IDs that you want in the response.

Main.Cloudfront.list_distributions_by_key_group2020_05_31 — Method

ListDistributionsByKeyGroup2020_05_31()

Gets a list of distribution IDs for distributions that have a cache behavior that references the specified key group. You can optionally specify the maximum number of items to receive in the response. If the total number of items in the list exceeds the maximum that you specify, or the default maximum, the response is paginated. To get the next page of items, send a subsequent request that specifies the NextMarker value from the current response as the Marker value in the subsequent request.

Required Parameters

KeyGroupId: The ID of the key group whose associated distribution IDs you are listing.

Optional Parameters

Marker: Use this field when paginating results to indicate where to begin in your list of distribution IDs. The response includes distribution IDs in the list that occur after the marker. To get the next page of the list, set this field’s value to the value of NextMarker from the current page’s response.

MaxItems: The maximum number of distribution IDs that you want in the response.

Main.Cloudfront.list_distributions_by_origin_request_policy_id2020_05_31 — Method

ListDistributionsByOriginRequestPolicyId2020_05_31()

Gets a list of distribution IDs for distributions that have a cache behavior that’s associated with the specified origin request policy. You can optionally specify the maximum number of items to receive in the response. If the total number of items in the list exceeds the maximum that you specify, or the default maximum, the response is paginated. To get the next page of items, send a subsequent request that specifies the NextMarker value from the current response as the Marker value in the subsequent request.

Required Parameters

OriginRequestPolicyId: The ID of the origin request policy whose associated distribution IDs you want to list.

Optional Parameters

Marker: Use this field when paginating results to indicate where to begin in your list of distribution IDs. The response includes distribution IDs in the list that occur after the marker. To get the next page of the list, set this field’s value to the value of NextMarker from the current page’s response.

MaxItems: The maximum number of distribution IDs that you want in the response.

Main.Cloudfront.list_distributions_by_realtime_log_config2020_05_31 — Method

ListDistributionsByRealtimeLogConfig2020_05_31()

Gets a list of distributions that have a cache behavior that’s associated with the specified real-time log configuration. You can specify the real-time log configuration by its name or its Amazon Resource Name (ARN). You must provide at least one. If you provide both, CloudFront uses the name to identify the real-time log configuration to list distributions for. You can optionally specify the maximum number of items to receive in the response. If the total number of items in the list exceeds the maximum that you specify, or the default maximum, the response is paginated. To get the next page of items, send a subsequent request that specifies the NextMarker value from the current response as the Marker value in the subsequent request.

Optional Parameters

Marker: Use this field when paginating results to indicate where to begin in your list of distributions. The response includes distributions in the list that occur after the marker. To get the next page of the list, set this field’s value to the value of NextMarker from the current page’s response.

MaxItems: The maximum number of distributions that you want in the response.

RealtimeLogConfigArn: The Amazon Resource Name (ARN) of the real-time log configuration whose associated distributions you want to list.

RealtimeLogConfigName: The name of the real-time log configuration whose associated distributions you want to list.

Main.Cloudfront.list_distributions_by_web_aclid2020_05_31 — Method

ListDistributionsByWebACLId2020_05_31()

List the distributions that are associated with a specified AWS WAF web ACL.

Required Parameters

WebACLId: The ID of the AWS WAF web ACL that you want to list the associated distributions. If you specify "null" for the ID, the request returns a list of the distributions that aren't associated with a web ACL.

Optional Parameters

Marker: Use Marker and MaxItems to control pagination of results. If you have more than MaxItems distributions that satisfy the request, the response includes a NextMarker element. To get the next page of results, submit another request. For the value of Marker, specify the value of NextMarker from the last response. (For the first request, omit Marker.)

MaxItems: The maximum number of distributions that you want CloudFront to return in the response body. The maximum and default values are both 100.

Main.Cloudfront.list_field_level_encryption_configs2020_05_31 — Method

ListFieldLevelEncryptionConfigs2020_05_31()

List all field-level encryption configurations that have been created in CloudFront for this account.

Optional Parameters

Marker: Use this when paginating results to indicate where to begin in your list of configurations. The results include configurations in the list that occur after the marker. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response (which is also the ID of the last configuration on that page).

MaxItems: The maximum number of field-level encryption configurations you want in the response body.

Main.Cloudfront.list_field_level_encryption_profiles2020_05_31 — Method

ListFieldLevelEncryptionProfiles2020_05_31()

Request a list of field-level encryption profiles that have been created in CloudFront for this account.

Optional Parameters

Marker: Use this when paginating results to indicate where to begin in your list of profiles. The results include profiles in the list that occur after the marker. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response (which is also the ID of the last profile on that page).

MaxItems: The maximum number of field-level encryption profiles you want in the response body.

Main.Cloudfront.list_invalidations2020_05_31 — Method

ListInvalidations2020_05_31()

Lists invalidation batches.

Required Parameters

DistributionId: The distribution's ID.

Optional Parameters

Marker: Use this parameter when paginating results to indicate where to begin in your list of invalidation batches. Because the results are returned in decreasing order from most recent to oldest, the most recent results are on the first page, the second page will contain earlier results, and so on. To get the next page of results, set Marker to the value of the NextMarker from the current page's response. This value is the same as the ID of the last invalidation batch on that page.

MaxItems: The maximum number of invalidation batches that you want in the response body.

Main.Cloudfront.list_key_groups2020_05_31 — Method

ListKeyGroups2020_05_31()

Gets a list of key groups. You can optionally specify the maximum number of items to receive in the response. If the total number of items in the list exceeds the maximum that you specify, or the default maximum, the response is paginated. To get the next page of items, send a subsequent request that specifies the NextMarker value from the current response as the Marker value in the subsequent request.

Optional Parameters

Marker: Use this field when paginating results to indicate where to begin in your list of key groups. The response includes key groups in the list that occur after the marker. To get the next page of the list, set this field’s value to the value of NextMarker from the current page’s response.

MaxItems: The maximum number of key groups that you want in the response.

Main.Cloudfront.list_origin_request_policies2020_05_31 — Method

ListOriginRequestPolicies2020_05_31()

Gets a list of origin request policies. You can optionally apply a filter to return only the managed policies created by AWS, or only the custom policies created in your AWS account. You can optionally specify the maximum number of items to receive in the response. If the total number of items in the list exceeds the maximum that you specify, or the default maximum, the response is paginated. To get the next page of items, send a subsequent request that specifies the NextMarker value from the current response as the Marker value in the subsequent request.

Optional Parameters

Marker: Use this field when paginating results to indicate where to begin in your list of origin request policies. The response includes origin request policies in the list that occur after the marker. To get the next page of the list, set this field’s value to the value of NextMarker from the current page’s response.

MaxItems: The maximum number of origin request policies that you want in the response.

Type: A filter to return only the specified kinds of origin request policies. Valid values are: managed – Returns only the managed policies created by AWS. custom – Returns only the custom policies created in your AWS account.

Main.Cloudfront.list_public_keys2020_05_31 — Method

ListPublicKeys2020_05_31()

List all public keys that have been added to CloudFront for this account.

Optional Parameters

Marker: Use this when paginating results to indicate where to begin in your list of public keys. The results include public keys in the list that occur after the marker. To get the next page of results, set the Marker to the value of the NextMarker from the current page's response (which is also the ID of the last public key on that page).

MaxItems: The maximum number of public keys you want in the response body.

Main.Cloudfront.list_realtime_log_configs2020_05_31 — Method

ListRealtimeLogConfigs2020_05_31()

Gets a list of real-time log configurations. You can optionally specify the maximum number of items to receive in the response. If the total number of items in the list exceeds the maximum that you specify, or the default maximum, the response is paginated. To get the next page of items, send a subsequent request that specifies the NextMarker value from the current response as the Marker value in the subsequent request.

Optional Parameters

Marker: Use this field when paginating results to indicate where to begin in your list of real-time log configurations. The response includes real-time log configurations in the list that occur after the marker. To get the next page of the list, set this field’s value to the value of NextMarker from the current page’s response.

MaxItems: The maximum number of real-time log configurations that you want in the response.

Main.Cloudfront.list_streaming_distributions2020_05_31 — Method

ListStreamingDistributions2020_05_31()

List streaming distributions.

Optional Parameters

Marker: The value that you provided for the Marker request parameter.

MaxItems: The value that you provided for the MaxItems request parameter.

Main.Cloudfront.list_tags_for_resource2020_05_31 — Method

ListTagsForResource2020_05_31()

List tags for a CloudFront resource.

Required Parameters

Resource: An ARN of a CloudFront resource.

Main.Cloudfront.tag_resource2020_05_31 — Method

TagResource2020_05_31()

Add tags to a CloudFront resource.

Required Parameters

Resource: An ARN of a CloudFront resource.

Tags: A complex type that contains zero or more Tag elements.

Main.Cloudfront.untag_resource2020_05_31 — Method

UntagResource2020_05_31()

Remove tags from a CloudFront resource.

Required Parameters

Resource: An ARN of a CloudFront resource.

TagKeys: A complex type that contains zero or more Tag key elements.

Main.Cloudfront.update_cache_policy2020_05_31 — Method

UpdateCachePolicy2020_05_31()

Updates a cache policy configuration. When you update a cache policy configuration, all the fields are updated with the values provided in the request. You cannot update some fields independent of others. To update a cache policy configuration: Use GetCachePolicyConfig to get the current configuration. Locally modify the fields in the cache policy configuration that you want to update. Call UpdateCachePolicy by providing the entire cache policy configuration, including the fields that you modified and those that you didn’t.

Required Parameters

CachePolicyConfig: A cache policy configuration.

Id: The unique identifier for the cache policy that you are updating. The identifier is returned in a cache behavior’s CachePolicyId field in the response to GetDistributionConfig.

Optional Parameters

If-Match: The version of the cache policy that you are updating. The version is returned in the cache policy’s ETag field in the response to GetCachePolicyConfig.

Main.Cloudfront.update_cloud_front_origin_access_identity2020_05_31 — Method

UpdateCloudFrontOriginAccessIdentity2020_05_31()

Update an origin access identity.

Required Parameters

CloudFrontOriginAccessIdentityConfig: The identity's configuration information.

Id: The identity's id.

Optional Parameters

If-Match: The value of the ETag header that you received when retrieving the identity's configuration. For example: E2QWRUHAPOMQZL.

Main.Cloudfront.update_distribution2020_05_31 — Method

UpdateDistribution2020_05_31()

Updates the configuration for a web distribution. When you update a distribution, there are more required fields than when you create a distribution. When you update your distribution by using this API action, follow the steps here to get the current configuration and then make your updates, to make sure that you include all of the required fields. To view a summary, see Required Fields for Create Distribution and Update Distribution in the Amazon CloudFront Developer Guide. The update process includes getting the current distribution configuration, updating the XML document that is returned to make your changes, and then submitting an UpdateDistribution request to make the updates. For information about updating a distribution using the CloudFront console instead, see Creating a Distribution in the Amazon CloudFront Developer Guide. To update a web distribution using the CloudFront API Submit a GetDistributionConfig request to get the current configuration and an Etag header for the distribution. If you update the distribution again, you must get a new Etag header. Update the XML document that was returned in the response to your GetDistributionConfig request to include your changes. When you edit the XML file, be aware of the following: You must strip out the ETag parameter that is returned. Additional fields are required when you update a distribution. There may be fields included in the XML file for features that you haven't configured for your distribution. This is expected and required to successfully update the distribution. You can't change the value of CallerReference. If you try to change this value, CloudFront returns an IllegalUpdate error. The new configuration replaces the existing configuration; the values that you specify in an UpdateDistribution request are not merged into your existing configuration. When you add, delete, or replace values in an element that allows multiple values (for example, CNAME), you must specify all of the values that you want to appear in the updated distribution. In addition, you must update the corresponding Quantity element. Submit an UpdateDistribution request to update the configuration for your distribution: In the request body, include the XML document that you updated in Step 2. The request body must include an XML document with a DistributionConfig element. Set the value of the HTTP If-Match header to the value of the ETag header that CloudFront returned when you submitted the GetDistributionConfig request in Step 1. Review the response to the UpdateDistribution request to confirm that the configuration was successfully updated. Optional: Submit a GetDistribution request to confirm that your changes have propagated. When propagation is complete, the value of Status is Deployed.

Required Parameters

DistributionConfig: The distribution's configuration information.

Id: The distribution's id.

Optional Parameters

If-Match: The value of the ETag header that you received when retrieving the distribution's configuration. For example: E2QWRUHAPOMQZL.

Main.Cloudfront.update_field_level_encryption_config2020_05_31 — Method

UpdateFieldLevelEncryptionConfig2020_05_31()

Update a field-level encryption configuration.

Required Parameters

FieldLevelEncryptionConfig: Request to update a field-level encryption configuration.

Id: The ID of the configuration you want to update.

Optional Parameters

If-Match: The value of the ETag header that you received when retrieving the configuration identity to update. For example: E2QWRUHAPOMQZL.

Main.Cloudfront.update_field_level_encryption_profile2020_05_31 — Method

UpdateFieldLevelEncryptionProfile2020_05_31()

Update a field-level encryption profile.

Required Parameters

FieldLevelEncryptionProfileConfig: Request to update a field-level encryption profile.

Id: The ID of the field-level encryption profile request.

Optional Parameters

If-Match: The value of the ETag header that you received when retrieving the profile identity to update. For example: E2QWRUHAPOMQZL.

Main.Cloudfront.update_key_group2020_05_31 — Method

UpdateKeyGroup2020_05_31()

Updates a key group. When you update a key group, all the fields are updated with the values provided in the request. You cannot update some fields independent of others. To update a key group: Get the current key group with GetKeyGroup or GetKeyGroupConfig. Locally modify the fields in the key group that you want to update. For example, add or remove public key IDs. Call UpdateKeyGroup with the entire key group object, including the fields that you modified and those that you didn’t.

Required Parameters

Id: The identifier of the key group that you are updating.

KeyGroupConfig: The key group configuration.

Optional Parameters

If-Match: The version of the key group that you are updating. The version is the key group’s ETag value.

Main.Cloudfront.update_origin_request_policy2020_05_31 — Method

UpdateOriginRequestPolicy2020_05_31()

Updates an origin request policy configuration. When you update an origin request policy configuration, all the fields are updated with the values provided in the request. You cannot update some fields independent of others. To update an origin request policy configuration: Use GetOriginRequestPolicyConfig to get the current configuration. Locally modify the fields in the origin request policy configuration that you want to update. Call UpdateOriginRequestPolicy by providing the entire origin request policy configuration, including the fields that you modified and those that you didn’t.

Required Parameters

Id: The unique identifier for the origin request policy that you are updating. The identifier is returned in a cache behavior’s OriginRequestPolicyId field in the response to GetDistributionConfig.

OriginRequestPolicyConfig: An origin request policy configuration.

Optional Parameters

If-Match: The version of the origin request policy that you are updating. The version is returned in the origin request policy’s ETag field in the response to GetOriginRequestPolicyConfig.

Main.Cloudfront.update_public_key2020_05_31 — Method

UpdatePublicKey2020_05_31()

Update public key information. Note that the only value you can change is the comment.

Required Parameters

Id: The identifier of the public key that you are updating.

PublicKeyConfig: A public key configuration.

Optional Parameters

If-Match: The value of the ETag header that you received when retrieving the public key to update. For example: E2QWRUHAPOMQZL.

Main.Cloudfront.update_realtime_log_config2020_05_31 — Method

UpdateRealtimeLogConfig2020_05_31()

Updates a real-time log configuration. When you update a real-time log configuration, all the parameters are updated with the values provided in the request. You cannot update some parameters independent of others. To update a real-time log configuration: Call GetRealtimeLogConfig to get the current real-time log configuration. Locally modify the parameters in the real-time log configuration that you want to update. Call this API (UpdateRealtimeLogConfig) by providing the entire real-time log configuration, including the parameters that you modified and those that you didn’t. You cannot update a real-time log configuration’s Name or ARN.

Optional Parameters

ARN: The Amazon Resource Name (ARN) for this real-time log configuration.

EndPoints: Contains information about the Amazon Kinesis data stream where you are sending real-time log data.

Fields: A list of fields to include in each real-time log record. For more information about fields, see Real-time log configuration fields in the Amazon CloudFront Developer Guide.

Name: The name for this real-time log configuration.

SamplingRate: The sampling rate for this real-time log configuration. The sampling rate determines the percentage of viewer requests that are represented in the real-time log data. You must provide an integer between 1 and 100, inclusive.

Main.Cloudfront.update_streaming_distribution2020_05_31 — Method

UpdateStreamingDistribution2020_05_31()

Update a streaming distribution.

Required Parameters

Id: The streaming distribution's id.

StreamingDistributionConfig: The streaming distribution's configuration information.

Optional Parameters

If-Match: The value of the ETag header that you received when retrieving the streaming distribution's configuration. For example: E2QWRUHAPOMQZL.