Cloudsearch

This page documents function available when using the Cloudsearch module, created with @service Cloudsearch.

Index

Main.Cloudsearch.build_suggesters

Main.Cloudsearch.create_domain

Main.Cloudsearch.define_analysis_scheme

Main.Cloudsearch.define_expression

Main.Cloudsearch.define_index_field

Main.Cloudsearch.define_suggester

Main.Cloudsearch.delete_analysis_scheme

Main.Cloudsearch.delete_domain

Main.Cloudsearch.delete_expression

Main.Cloudsearch.delete_index_field

Main.Cloudsearch.delete_suggester

Main.Cloudsearch.describe_analysis_schemes

Main.Cloudsearch.describe_availability_options

Main.Cloudsearch.describe_domain_endpoint_options

Main.Cloudsearch.describe_domains

Main.Cloudsearch.describe_expressions

Main.Cloudsearch.describe_index_fields

Main.Cloudsearch.describe_scaling_parameters

Main.Cloudsearch.describe_service_access_policies

Main.Cloudsearch.describe_suggesters

Main.Cloudsearch.index_documents

Main.Cloudsearch.list_domain_names

Main.Cloudsearch.update_availability_options

Main.Cloudsearch.update_domain_endpoint_options

Main.Cloudsearch.update_scaling_parameters

Main.Cloudsearch.update_service_access_policies

Documentation

Main.Cloudsearch.build_suggesters — Function

build_suggesters(domain_name)
build_suggesters(domain_name, params::Dict{String,<:Any})

Indexes the search suggestions. For more information, see Configuring Suggesters in the Amazon CloudSearch Developer Guide.

Arguments

domain_name:

Main.Cloudsearch.create_domain — Function

create_domain(domain_name)
create_domain(domain_name, params::Dict{String,<:Any})

Creates a new search domain. For more information, see <a href="http://docs.aws.amazon.com/cloudsearch/latest/developerguide/creating-domains.html" target="_blank">Creating a Search Domain</a> in the Amazon CloudSearch Developer Guide.

Arguments

domain_name: A name for the domain you are creating. Allowed characters are a-z (lower- case letters), 0-9, and hyphen (-). Domain names must start with a letter or number and be at least 3 and no more than 28 characters long.

Main.Cloudsearch.define_analysis_scheme — Function

define_analysis_scheme(analysis_scheme, domain_name)
define_analysis_scheme(analysis_scheme, domain_name, params::Dict{String,<:Any})

Configures an analysis scheme that can be applied to a text or text-array field to define language-specific text processing options. For more information, see <a href="http://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-analysis- schemes.html" target="_blank">Configuring Analysis Schemes</a> in the Amazon CloudSearch Developer Guide.

Arguments

analysis_scheme:

domain_name:

Main.Cloudsearch.define_expression — Function

define_expression(domain_name, expression)
define_expression(domain_name, expression, params::Dict{String,<:Any})

Configures an $Expression$ for the search domain. Used to create new expressions and modify existing ones. If the expression exists, the new configuration replaces the old one. For more information, see <a href="http://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring- expressions.html" target="_blank">Configuring Expressions</a> in the Amazon CloudSearch Developer Guide.

Arguments

domain_name:

expression:

Main.Cloudsearch.define_index_field — Function

define_index_field(domain_name, index_field)
define_index_field(domain_name, index_field, params::Dict{String,<:Any})

Configures an $IndexField$ for the search domain. Used to create new fields and modify existing ones. You must specify the name of the domain you are configuring and an index field configuration. The index field configuration specifies a unique name, the index field type, and the options you want to configure for the field. The options you can specify depend on the $IndexFieldType$. If the field exists, the new configuration replaces the old one. For more information, see <a href="http://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-index- fields.html" target="_blank">Configuring Index Fields</a> in the Amazon CloudSearch Developer Guide.

Arguments

domain_name:

index_field: The index field and field options you want to configure.

Main.Cloudsearch.define_suggester — Function

define_suggester(domain_name, suggester)
define_suggester(domain_name, suggester, params::Dict{String,<:Any})

Configures a suggester for a domain. A suggester enables you to display possible matches before users finish typing their queries. When you configure a suggester, you must specify the name of the text field you want to search for possible matches and a unique name for the suggester. For more information, see <a href="http://docs.aws.amazon.com/cloudsearch/latest/developerguide/getting-suggestions.html" target="_blank">Getting Search Suggestions</a> in the Amazon CloudSearch Developer Guide.

Arguments

domain_name:

suggester:

Main.Cloudsearch.delete_analysis_scheme — Function

delete_analysis_scheme(analysis_scheme_name, domain_name)
delete_analysis_scheme(analysis_scheme_name, domain_name, params::Dict{String,<:Any})

Deletes an analysis scheme. For more information, see <a href="http://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-analysis- schemes.html" target="_blank">Configuring Analysis Schemes</a> in the Amazon CloudSearch Developer Guide.

Arguments

analysis_scheme_name: The name of the analysis scheme you want to delete.

domain_name:

Main.Cloudsearch.delete_domain — Function

delete_domain(domain_name)
delete_domain(domain_name, params::Dict{String,<:Any})

Permanently deletes a search domain and all of its data. Once a domain has been deleted, it cannot be recovered. For more information, see <a href="http://docs.aws.amazon.com/cloudsearch/latest/developerguide/deleting-domains.html" target="_blank">Deleting a Search Domain</a> in the Amazon CloudSearch Developer Guide.

Arguments

domain_name: The name of the domain you want to permanently delete.

Main.Cloudsearch.delete_expression — Function

delete_expression(domain_name, expression_name)
delete_expression(domain_name, expression_name, params::Dict{String,<:Any})

Removes an $Expression$ from the search domain. For more information, see <a href="http://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring- expressions.html" target="_blank">Configuring Expressions</a> in the Amazon CloudSearch Developer Guide.

Arguments

domain_name:

expression_name: The name of the $Expression$ to delete.

Main.Cloudsearch.delete_index_field — Function

delete_index_field(domain_name, index_field_name)
delete_index_field(domain_name, index_field_name, params::Dict{String,<:Any})

Removes an $IndexField$ from the search domain. For more information, see <a href="http://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-index- fields.html" target="_blank">Configuring Index Fields</a> in the Amazon CloudSearch Developer Guide.

Arguments

domain_name:

index_field_name: The name of the index field your want to remove from the domain's indexing options.

Main.Cloudsearch.delete_suggester — Function

delete_suggester(domain_name, suggester_name)
delete_suggester(domain_name, suggester_name, params::Dict{String,<:Any})

Deletes a suggester. For more information, see <a href="http://docs.aws.amazon.com/cloudsearch/latest/developerguide/getting-suggestions.html" target="_blank">Getting Search Suggestions</a> in the Amazon CloudSearch Developer Guide.

Arguments

domain_name:

suggester_name: Specifies the name of the suggester you want to delete.

Main.Cloudsearch.describe_analysis_schemes — Function

describe_analysis_schemes(domain_name)
describe_analysis_schemes(domain_name, params::Dict{String,<:Any})

Gets the analysis schemes configured for a domain. An analysis scheme defines language- specific text processing options for a text field. Can be limited to specific analysis schemes by name. By default, shows all analysis schemes and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see <a href="http://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-analysis- schemes.html" target="_blank">Configuring Analysis Schemes</a> in the Amazon CloudSearch Developer Guide.

Arguments

domain_name: The name of the domain you want to describe.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AnalysisSchemeNames": The analysis schemes you want to describe.

"Deployed": Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

Main.Cloudsearch.describe_availability_options — Function

describe_availability_options(domain_name)
describe_availability_options(domain_name, params::Dict{String,<:Any})

Gets the availability options configured for a domain. By default, shows the configuration with any pending changes. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see <a href="http://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-availability- options.html" target="_blank">Configuring Availability Options</a> in the Amazon CloudSearch Developer Guide.

Arguments

domain_name: The name of the domain you want to describe.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Deployed": Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

Main.Cloudsearch.describe_domain_endpoint_options — Function

describe_domain_endpoint_options(domain_name)
describe_domain_endpoint_options(domain_name, params::Dict{String,<:Any})

Returns the domain's endpoint options, specifically whether all requests to the domain must arrive over HTTPS. For more information, see <a href="http://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-domain- endpoint-options.html" target="_blank">Configuring Domain Endpoint Options</a> in the Amazon CloudSearch Developer Guide.

Arguments

domain_name: A string that represents the name of a domain.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Deployed": Whether to retrieve the latest configuration (which might be in a Processing state) or the current, active configuration. Defaults to false.

Main.Cloudsearch.describe_domains — Function

describe_domains()
describe_domains(params::Dict{String,<:Any})

Gets information about the search domains owned by this account. Can be limited to specific domains. Shows all domains by default. To get the number of searchable documents in a domain, use the console or submit a matchall request to your domain's search endpoint: q=matchall&amp;q.parser=structured&amp;size=0. For more information, see <a href="http://docs.aws.amazon.com/cloudsearch/latest/developerguide/getting-domain-info.html" target="_blank">Getting Information about a Search Domain</a> in the Amazon CloudSearch Developer Guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DomainNames": The names of the domains you want to include in the response.

Main.Cloudsearch.describe_expressions — Function

describe_expressions(domain_name)
describe_expressions(domain_name, params::Dict{String,<:Any})

Gets the expressions configured for the search domain. Can be limited to specific expressions by name. By default, shows all expressions and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see <a href="http://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring- expressions.html" target="_blank">Configuring Expressions</a> in the Amazon CloudSearch Developer Guide.

Arguments

domain_name: The name of the domain you want to describe.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Deployed": Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

"ExpressionNames": Limits the $DescribeExpressions$ response to the specified expressions. If not specified, all expressions are shown.

Main.Cloudsearch.describe_index_fields — Function

describe_index_fields(domain_name)
describe_index_fields(domain_name, params::Dict{String,<:Any})

Gets information about the index fields configured for the search domain. Can be limited to specific fields by name. By default, shows all fields and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see <a href="http://docs.aws.amazon.com/cloudsearch/latest/developerguide/getting-domain-info.html" target="_blank">Getting Domain Information</a> in the Amazon CloudSearch Developer Guide.

Arguments

domain_name: The name of the domain you want to describe.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Deployed": Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

"FieldNames": A list of the index fields you want to describe. If not specified, information is returned for all configured index fields.

Main.Cloudsearch.describe_scaling_parameters — Function

describe_scaling_parameters(domain_name)
describe_scaling_parameters(domain_name, params::Dict{String,<:Any})

Gets the scaling parameters configured for a domain. A domain's scaling parameters specify the desired search instance type and replication count. For more information, see <a href="http://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-scaling- options.html" target="_blank">Configuring Scaling Options</a> in the Amazon CloudSearch Developer Guide.

Arguments

domain_name:

Main.Cloudsearch.describe_service_access_policies — Function

describe_service_access_policies(domain_name)
describe_service_access_policies(domain_name, params::Dict{String,<:Any})

Gets information about the access policies that control access to the domain's document and search endpoints. By default, shows the configuration with any pending changes. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see <a href="http://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-access.html" target="_blank">Configuring Access for a Search Domain</a> in the Amazon CloudSearch Developer Guide.

Arguments

domain_name: The name of the domain you want to describe.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Deployed": Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

Main.Cloudsearch.describe_suggesters — Function

describe_suggesters(domain_name)
describe_suggesters(domain_name, params::Dict{String,<:Any})

Gets the suggesters configured for a domain. A suggester enables you to display possible matches before users finish typing their queries. Can be limited to specific suggesters by name. By default, shows all suggesters and includes any pending changes to the configuration. Set the Deployed option to true to show the active configuration and exclude pending changes. For more information, see <a href="http://docs.aws.amazon.com/cloudsearch/latest/developerguide/getting-suggestions.html" target="_blank">Getting Search Suggestions</a> in the Amazon CloudSearch Developer Guide.

Arguments

domain_name: The name of the domain you want to describe.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Deployed": Whether to display the deployed configuration (true) or include any pending changes (false). Defaults to false.

"SuggesterNames": The suggesters you want to describe.

Main.Cloudsearch.index_documents — Function

index_documents(domain_name)
index_documents(domain_name, params::Dict{String,<:Any})

Tells the search domain to start indexing its documents using the latest indexing options. This operation must be invoked to activate options whose OptionStatus is RequiresIndexDocuments.

Arguments

domain_name:

Main.Cloudsearch.list_domain_names — Function

list_domain_names()
list_domain_names(params::Dict{String,<:Any})

Lists all search domains owned by an account.

Main.Cloudsearch.update_availability_options — Function

update_availability_options(domain_name, multi_az)
update_availability_options(domain_name, multi_az, params::Dict{String,<:Any})

Configures the availability options for a domain. Enabling the Multi-AZ option expands an Amazon CloudSearch domain to an additional Availability Zone in the same Region to increase fault tolerance in the event of a service disruption. Changes to the Multi-AZ option can take about half an hour to become active. For more information, see <a href="http://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-availability- options.html" target="_blank">Configuring Availability Options</a> in the Amazon CloudSearch Developer Guide.

Arguments

domain_name:

multi_az: You expand an existing search domain to a second Availability Zone by setting the Multi-AZ option to true. Similarly, you can turn off the Multi-AZ option to downgrade the domain to a single Availability Zone by setting the Multi-AZ option to false.

Main.Cloudsearch.update_domain_endpoint_options — Function

update_domain_endpoint_options(domain_endpoint_options, domain_name)
update_domain_endpoint_options(domain_endpoint_options, domain_name, params::Dict{String,<:Any})

Updates the domain's endpoint options, specifically whether all requests to the domain must arrive over HTTPS. For more information, see <a href="http://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-domain- endpoint-options.html" target="_blank">Configuring Domain Endpoint Options</a> in the Amazon CloudSearch Developer Guide.

Arguments

domain_endpoint_options: Whether to require that all requests to the domain arrive over HTTPS. We recommend Policy-Min-TLS-1-2-2019-07 for TLSSecurityPolicy. For compatibility with older clients, the default is Policy-Min-TLS-1-0-2019-07.

domain_name: A string that represents the name of a domain.

Main.Cloudsearch.update_scaling_parameters — Function

update_scaling_parameters(domain_name, scaling_parameters)
update_scaling_parameters(domain_name, scaling_parameters, params::Dict{String,<:Any})

Configures scaling parameters for a domain. A domain's scaling parameters specify the desired search instance type and replication count. Amazon CloudSearch will still automatically scale your domain based on the volume of data and traffic, but not below the desired instance type and replication count. If the Multi-AZ option is enabled, these values control the resources used per Availability Zone. For more information, see <a href="http://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-scaling- options.html" target="_blank">Configuring Scaling Options</a> in the Amazon CloudSearch Developer Guide.

Arguments

domain_name:

scaling_parameters:

Main.Cloudsearch.update_service_access_policies — Function

update_service_access_policies(access_policies, domain_name)
update_service_access_policies(access_policies, domain_name, params::Dict{String,<:Any})

Configures the access rules that control access to the domain's document and search endpoints. For more information, see <a href="http://docs.aws.amazon.com/cloudsearch/latest/developerguide/configuring-access.html" target="_blank"> Configuring Access for an Amazon CloudSearch Domain</a>.

Arguments

access_policies: The access rules you want to configure. These rules replace any existing rules.

domain_name: