Cloudwatch Logs

This page documents function available when using the Cloudwatch_Logs module, created with @service Cloudwatch_Logs.

Index

Main.Cloudwatch_Logs.associate_kms_key

Main.Cloudwatch_Logs.cancel_export_task

Main.Cloudwatch_Logs.create_export_task

Main.Cloudwatch_Logs.create_log_group

Main.Cloudwatch_Logs.create_log_stream

Main.Cloudwatch_Logs.delete_destination

Main.Cloudwatch_Logs.delete_log_group

Main.Cloudwatch_Logs.delete_log_stream

Main.Cloudwatch_Logs.delete_metric_filter

Main.Cloudwatch_Logs.delete_query_definition

Main.Cloudwatch_Logs.delete_resource_policy

Main.Cloudwatch_Logs.delete_retention_policy

Main.Cloudwatch_Logs.delete_subscription_filter

Main.Cloudwatch_Logs.describe_destinations

Main.Cloudwatch_Logs.describe_export_tasks

Main.Cloudwatch_Logs.describe_log_groups

Main.Cloudwatch_Logs.describe_log_streams

Main.Cloudwatch_Logs.describe_metric_filters

Main.Cloudwatch_Logs.describe_queries

Main.Cloudwatch_Logs.describe_query_definitions

Main.Cloudwatch_Logs.describe_resource_policies

Main.Cloudwatch_Logs.describe_subscription_filters

Main.Cloudwatch_Logs.disassociate_kms_key

Main.Cloudwatch_Logs.filter_log_events

Main.Cloudwatch_Logs.get_log_events

Main.Cloudwatch_Logs.get_log_group_fields

Main.Cloudwatch_Logs.get_log_record

Main.Cloudwatch_Logs.get_query_results

Main.Cloudwatch_Logs.list_tags_log_group

Main.Cloudwatch_Logs.put_destination

Main.Cloudwatch_Logs.put_destination_policy

Main.Cloudwatch_Logs.put_log_events

Main.Cloudwatch_Logs.put_metric_filter

Main.Cloudwatch_Logs.put_query_definition

Main.Cloudwatch_Logs.put_resource_policy

Main.Cloudwatch_Logs.put_retention_policy

Main.Cloudwatch_Logs.put_subscription_filter

Main.Cloudwatch_Logs.start_query

Main.Cloudwatch_Logs.stop_query

Main.Cloudwatch_Logs.tag_log_group

Main.Cloudwatch_Logs.test_metric_filter

Main.Cloudwatch_Logs.untag_log_group

Documentation

Main.Cloudwatch_Logs.associate_kms_key — Method

associate_kms_key(kms_key_id, log_group_name)
associate_kms_key(kms_key_id, log_group_name, params::Dict{String,<:Any})

Associates the specified AWS Key Management Service (AWS KMS) customer master key (CMK) with the specified log group. Associating an AWS KMS CMK with a log group overrides any existing associations between the log group and a CMK. After a CMK is associated with a log group, all newly ingested data for the log group is encrypted using the CMK. This association is stored as long as the data encrypted with the CMK is still within Amazon CloudWatch Logs. This enables Amazon CloudWatch Logs to decrypt this data whenever it is requested. CloudWatch Logs supports only symmetric CMKs. Do not use an associate an asymmetric CMK with your log group. For more information, see Using Symmetric and Asymmetric Keys. It can take up to 5 minutes for this operation to take effect. If you attempt to associate a CMK with a log group but the CMK does not exist or the CMK is disabled, you receive an InvalidParameterException error.

Arguments

kms_key_id: The Amazon Resource Name (ARN) of the CMK to use when encrypting log data. This must be a symmetric CMK. For more information, see Amazon Resource Names - AWS Key Management Service (AWS KMS) and Using Symmetric and Asymmetric Keys.

log_group_name: The name of the log group.

Main.Cloudwatch_Logs.cancel_export_task — Method

cancel_export_task(task_id)
cancel_export_task(task_id, params::Dict{String,<:Any})

Cancels the specified export task. The task must be in the PENDING or RUNNING state.

Arguments

task_id: The ID of the export task.

Main.Cloudwatch_Logs.create_export_task — Method

create_export_task(destination, from, log_group_name, to)
create_export_task(destination, from, log_group_name, to, params::Dict{String,<:Any})

Creates an export task, which allows you to efficiently export data from a log group to an Amazon S3 bucket. When you perform a CreateExportTask operation, you must use credentials that have permission to write to the S3 bucket that you specify as the destination. This is an asynchronous call. If all the required information is provided, this operation initiates an export task and responds with the ID of the task. After the task has started, you can use DescribeExportTasks to get the status of the export task. Each account can only have one active (RUNNING or PENDING) export task at a time. To cancel an export task, use CancelExportTask. You can export logs from multiple log groups or multiple time ranges to the same S3 bucket. To separate out log data for each export task, you can specify a prefix to be used as the Amazon S3 key prefix for all exported objects. Exporting to S3 buckets that are encrypted with AES-256 is supported. Exporting to S3 buckets encrypted with SSE-KMS is not supported.

Arguments

destination: The name of S3 bucket for the exported log data. The bucket must be in the same AWS region.

from: The start time of the range for the request, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp earlier than this time are not exported.

log_group_name: The name of the log group.

to: The end time of the range for the request, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp later than this time are not exported.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"destinationPrefix": The prefix used as the start of the key for every object exported. If you don't specify a value, the default is exportedlogs.

"logStreamNamePrefix": Export only log streams that match the provided prefix. If you don't specify a value, no prefix filter is applied.

"taskName": The name of the export task.

Main.Cloudwatch_Logs.create_log_group — Method

create_log_group(log_group_name)
create_log_group(log_group_name, params::Dict{String,<:Any})

Creates a log group with the specified name. You can create up to 20,000 log groups per account. You must use the following guidelines when naming a log group: Log group names must be unique within a region for an AWS account. Log group names can be between 1 and 512 characters long. Log group names consist of the following characters: a-z, A-Z, 0-9, '_' (underscore), '-' (hyphen), '/' (forward slash), '.' (period), and '#' (number sign) When you create a log group, by default the log events in the log group never expire. To set a retention policy so that events expire and are deleted after a specified time, use PutRetentionPolicy. If you associate a AWS Key Management Service (AWS KMS) customer master key (CMK) with the log group, ingested data is encrypted using the CMK. This association is stored as long as the data encrypted with the CMK is still within Amazon CloudWatch Logs. This enables Amazon CloudWatch Logs to decrypt this data whenever it is requested. If you attempt to associate a CMK with the log group but the CMK does not exist or the CMK is disabled, you receive an InvalidParameterException error. CloudWatch Logs supports only symmetric CMKs. Do not associate an asymmetric CMK with your log group. For more information, see Using Symmetric and Asymmetric Keys.

Arguments

log_group_name: The name of the log group.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"kmsKeyId": The Amazon Resource Name (ARN) of the CMK to use when encrypting log data. For more information, see Amazon Resource Names - AWS Key Management Service (AWS KMS).

"tags": The key-value pairs to use for the tags.

Main.Cloudwatch_Logs.create_log_stream — Method

create_log_stream(log_group_name, log_stream_name)
create_log_stream(log_group_name, log_stream_name, params::Dict{String,<:Any})

Creates a log stream for the specified log group. A log stream is a sequence of log events that originate from a single source, such as an application instance or a resource that is being monitored. There is no limit on the number of log streams that you can create for a log group. There is a limit of 50 TPS on CreateLogStream operations, after which transactions are throttled. You must use the following guidelines when naming a log stream: Log stream names must be unique within the log group. Log stream names can be between 1 and 512 characters long. The ':' (colon) and '*' (asterisk) characters are not allowed.

Arguments

log_group_name: The name of the log group.

log_stream_name: The name of the log stream.

Main.Cloudwatch_Logs.delete_destination — Method

delete_destination(destination_name)
delete_destination(destination_name, params::Dict{String,<:Any})

Deletes the specified destination, and eventually disables all the subscription filters that publish to it. This operation does not delete the physical resource encapsulated by the destination.

Arguments

destination_name: The name of the destination.

Main.Cloudwatch_Logs.delete_log_group — Method

delete_log_group(log_group_name)
delete_log_group(log_group_name, params::Dict{String,<:Any})

Deletes the specified log group and permanently deletes all the archived log events associated with the log group.

Arguments

log_group_name: The name of the log group.

Main.Cloudwatch_Logs.delete_log_stream — Method

delete_log_stream(log_group_name, log_stream_name)
delete_log_stream(log_group_name, log_stream_name, params::Dict{String,<:Any})

Deletes the specified log stream and permanently deletes all the archived log events associated with the log stream.

Arguments

log_group_name: The name of the log group.

log_stream_name: The name of the log stream.

Main.Cloudwatch_Logs.delete_metric_filter — Method

delete_metric_filter(filter_name, log_group_name)
delete_metric_filter(filter_name, log_group_name, params::Dict{String,<:Any})

Deletes the specified metric filter.

Arguments

filter_name: The name of the metric filter.

log_group_name: The name of the log group.

Main.Cloudwatch_Logs.delete_query_definition — Method

delete_query_definition(query_definition_id)
delete_query_definition(query_definition_id, params::Dict{String,<:Any})

Deletes a saved CloudWatch Logs Insights query definition. A query definition contains details about a saved CloudWatch Logs Insights query. Each DeleteQueryDefinition operation can delete one query definition. You must have the logs:DeleteQueryDefinition permission to be able to perform this operation.

Arguments

query_definition_id: The ID of the query definition that you want to delete. You can use DescribeQueryDefinitions to retrieve the IDs of your saved query definitions.

Main.Cloudwatch_Logs.delete_resource_policy — Method

delete_resource_policy()
delete_resource_policy(params::Dict{String,<:Any})

Deletes a resource policy from this account. This revokes the access of the identities in that policy to put log events to this account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"policyName": The name of the policy to be revoked. This parameter is required.

Main.Cloudwatch_Logs.delete_retention_policy — Method

delete_retention_policy(log_group_name)
delete_retention_policy(log_group_name, params::Dict{String,<:Any})

Deletes the specified retention policy. Log events do not expire if they belong to log groups without a retention policy.

Arguments

log_group_name: The name of the log group.

Main.Cloudwatch_Logs.delete_subscription_filter — Method

delete_subscription_filter(filter_name, log_group_name)
delete_subscription_filter(filter_name, log_group_name, params::Dict{String,<:Any})

Deletes the specified subscription filter.

Arguments

filter_name: The name of the subscription filter.

log_group_name: The name of the log group.

Main.Cloudwatch_Logs.describe_destinations — Method

describe_destinations()
describe_destinations(params::Dict{String,<:Any})

Lists all your destinations. The results are ASCII-sorted by destination name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DestinationNamePrefix": The prefix to match. If you don't specify a value, no prefix filter is applied.

"limit": The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

"nextToken": The token for the next set of items to return. (You received this token from a previous call.)

Main.Cloudwatch_Logs.describe_export_tasks — Method

describe_export_tasks()
describe_export_tasks(params::Dict{String,<:Any})

Lists the specified export tasks. You can list all your export tasks or filter the results based on task ID or task status.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"limit": The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

"nextToken": The token for the next set of items to return. (You received this token from a previous call.)

"statusCode": The status code of the export task. Specifying a status code filters the results to zero or more export tasks.

"taskId": The ID of the export task. Specifying a task ID filters the results to zero or one export tasks.

Main.Cloudwatch_Logs.describe_log_groups — Method

describe_log_groups()
describe_log_groups(params::Dict{String,<:Any})

Lists the specified log groups. You can list all your log groups or filter the results by prefix. The results are ASCII-sorted by log group name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"limit": The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

"logGroupNamePrefix": The prefix to match.

"nextToken": The token for the next set of items to return. (You received this token from a previous call.)

Main.Cloudwatch_Logs.describe_log_streams — Method

describe_log_streams(log_group_name)
describe_log_streams(log_group_name, params::Dict{String,<:Any})

Lists the log streams for the specified log group. You can list all the log streams or filter the results by prefix. You can also control how the results are ordered. This operation has a limit of five transactions per second, after which transactions are throttled.

Arguments

log_group_name: The name of the log group.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"descending": If the value is true, results are returned in descending order. If the value is to false, results are returned in ascending order. The default value is false.

"limit": The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

"logStreamNamePrefix": The prefix to match. If orderBy is LastEventTime, you cannot specify this parameter.

"nextToken": The token for the next set of items to return. (You received this token from a previous call.)

"orderBy": If the value is LogStreamName, the results are ordered by log stream name. If the value is LastEventTime, the results are ordered by the event time. The default value is LogStreamName. If you order the results by event time, you cannot specify the logStreamNamePrefix parameter. lastEventTimeStamp represents the time of the most recent log event in the log stream in CloudWatch Logs. This number is expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. lastEventTimeStamp updates on an eventual consistency basis. It typically updates in less than an hour from ingestion, but in rare situations might take longer.

Main.Cloudwatch_Logs.describe_metric_filters — Method

describe_metric_filters()
describe_metric_filters(params::Dict{String,<:Any})

Lists the specified metric filters. You can list all of the metric filters or filter the results by log name, prefix, metric name, or metric namespace. The results are ASCII-sorted by filter name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filterNamePrefix": The prefix to match. CloudWatch Logs uses the value you set here only if you also include the logGroupName parameter in your request.

"limit": The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

"logGroupName": The name of the log group.

"metricName": Filters results to include only those with the specified metric name. If you include this parameter in your request, you must also include the metricNamespace parameter.

"metricNamespace": Filters results to include only those in the specified namespace. If you include this parameter in your request, you must also include the metricName parameter.

"nextToken": The token for the next set of items to return. (You received this token from a previous call.)

Main.Cloudwatch_Logs.describe_queries — Method

describe_queries()
describe_queries(params::Dict{String,<:Any})

Returns a list of CloudWatch Logs Insights queries that are scheduled, executing, or have been executed recently in this account. You can request all queries or limit it to queries of a specific log group or queries with a certain status.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"logGroupName": Limits the returned queries to only those for the specified log group.

"maxResults": Limits the number of returned queries to the specified number.

"nextToken":

"status": Limits the returned queries to only those that have the specified status. Valid values are Cancelled, Complete, Failed, Running, and Scheduled.

Main.Cloudwatch_Logs.describe_query_definitions — Method

describe_query_definitions()
describe_query_definitions(params::Dict{String,<:Any})

This operation returns a paginated list of your saved CloudWatch Logs Insights query definitions. You can use the queryDefinitionNamePrefix parameter to limit the results to only the query definitions that have names that start with a certain string.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Limits the number of returned query definitions to the specified number.

"nextToken":

"queryDefinitionNamePrefix": Use this parameter to filter your results to only the query definitions that have names that start with the prefix you specify.

Main.Cloudwatch_Logs.describe_resource_policies — Method

describe_resource_policies()
describe_resource_policies(params::Dict{String,<:Any})

Lists the resource policies in this account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"limit": The maximum number of resource policies to be displayed with one call of this API.

"nextToken":

Main.Cloudwatch_Logs.describe_subscription_filters — Method

describe_subscription_filters(log_group_name)
describe_subscription_filters(log_group_name, params::Dict{String,<:Any})

Lists the subscription filters for the specified log group. You can list all the subscription filters or filter the results by prefix. The results are ASCII-sorted by filter name.

Arguments

log_group_name: The name of the log group.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filterNamePrefix": The prefix to match. If you don't specify a value, no prefix filter is applied.

"limit": The maximum number of items returned. If you don't specify a value, the default is up to 50 items.

"nextToken": The token for the next set of items to return. (You received this token from a previous call.)

Main.Cloudwatch_Logs.disassociate_kms_key — Method

disassociate_kms_key(log_group_name)
disassociate_kms_key(log_group_name, params::Dict{String,<:Any})

Disassociates the associated AWS Key Management Service (AWS KMS) customer master key (CMK) from the specified log group. After the AWS KMS CMK is disassociated from the log group, AWS CloudWatch Logs stops encrypting newly ingested data for the log group. All previously ingested data remains encrypted, and AWS CloudWatch Logs requires permissions for the CMK whenever the encrypted data is requested. Note that it can take up to 5 minutes for this operation to take effect.

Arguments

log_group_name: The name of the log group.

Main.Cloudwatch_Logs.filter_log_events — Method

filter_log_events(log_group_name)
filter_log_events(log_group_name, params::Dict{String,<:Any})

Lists log events from the specified log group. You can list all the log events or filter the results using a filter pattern, a time range, and the name of the log stream. By default, this operation returns as many log events as can fit in 1 MB (up to 10,000 log events) or all the events found within the time range that you specify. If the results include a token, then there are more log events available, and you can get additional results by specifying the token in a subsequent call. This operation can return empty results while there are more log events available through the token. The returned log events are sorted by event timestamp, the timestamp when the event was ingested by CloudWatch Logs, and the ID of the PutLogEvents request.

Arguments

log_group_name: The name of the log group to search.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"endTime": The end of the time range, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp later than this time are not returned.

"filterPattern": The filter pattern to use. For more information, see Filter and Pattern Syntax. If not provided, all the events are matched.

"interleaved": If the value is true, the operation makes a best effort to provide responses that contain events from multiple log streams within the log group, interleaved in a single response. If the value is false, all the matched log events in the first log stream are searched first, then those in the next log stream, and so on. The default is false. Important: Starting on June 17, 2019, this parameter is ignored and the value is assumed to be true. The response from this operation always interleaves events from multiple log streams within a log group.

"limit": The maximum number of events to return. The default is 10,000 events.

"logStreamNamePrefix": Filters the results to include only events from log streams that have names starting with this prefix. If you specify a value for both logStreamNamePrefix and logStreamNames, but the value for logStreamNamePrefix does not match any log stream names specified in logStreamNames, the action returns an InvalidParameterException error.

"logStreamNames": Filters the results to only logs from the log streams in this list. If you specify a value for both logStreamNamePrefix and logStreamNames, the action returns an InvalidParameterException error.

"nextToken": The token for the next set of events to return. (You received this token from a previous call.)

"startTime": The start of the time range, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp before this time are not returned. If you omit startTime and endTime the most recent log events are retrieved, to up 1 MB or 10,000 log events.

Main.Cloudwatch_Logs.get_log_events — Method

get_log_events(log_group_name, log_stream_name)
get_log_events(log_group_name, log_stream_name, params::Dict{String,<:Any})

Lists log events from the specified log stream. You can list all of the log events or filter using a time range. By default, this operation returns as many log events as can fit in a response size of 1MB (up to 10,000 log events). You can get additional log events by specifying one of the tokens in a subsequent call. This operation can return empty results while there are more log events available through the token.

Arguments

log_group_name: The name of the log group.

log_stream_name: The name of the log stream.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"endTime": The end of the time range, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp equal to or later than this time are not included.

"limit": The maximum number of log events returned. If you don't specify a value, the maximum is as many log events as can fit in a response size of 1 MB, up to 10,000 log events.

"nextToken": The token for the next set of items to return. (You received this token from a previous call.) Using this token works only when you specify true for startFromHead.

"startFromHead": If the value is true, the earliest log events are returned first. If the value is false, the latest log events are returned first. The default value is false. If you are using nextToken in this operation, you must specify true for startFromHead.

"startTime": The start of the time range, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. Events with a timestamp equal to this time or later than this time are included. Events with a timestamp earlier than this time are not included.

Main.Cloudwatch_Logs.get_log_group_fields — Method

get_log_group_fields(log_group_name)
get_log_group_fields(log_group_name, params::Dict{String,<:Any})

Returns a list of the fields that are included in log events in the specified log group, along with the percentage of log events that contain each field. The search is limited to a time period that you specify. In the results, fields that start with @ are fields generated by CloudWatch Logs. For example, @timestamp is the timestamp of each log event. For more information about the fields that are generated by CloudWatch logs, see Supported Logs and Discovered Fields. The response results are sorted by the frequency percentage, starting with the highest percentage.

Arguments

log_group_name: The name of the log group to search.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"time": The time to set as the center of the query. If you specify time, the 8 minutes before and 8 minutes after this time are searched. If you omit time, the past 15 minutes are queried. The time value is specified as epoch time, the number of seconds since January 1, 1970, 00:00:00 UTC.

Main.Cloudwatch_Logs.get_log_record — Method

get_log_record(log_record_pointer)
get_log_record(log_record_pointer, params::Dict{String,<:Any})

Retrieves all of the fields and values of a single log event. All fields are retrieved, even if the original query that produced the logRecordPointer retrieved only a subset of fields. Fields are returned as field name/field value pairs. The full unparsed log event is returned within @message.

Arguments

log_record_pointer: The pointer corresponding to the log event record you want to retrieve. You get this from the response of a GetQueryResults operation. In that response, the value of the @ptr field for a log event is the value to use as logRecordPointer to retrieve that complete log event record.

Main.Cloudwatch_Logs.get_query_results — Method

get_query_results(query_id)
get_query_results(query_id, params::Dict{String,<:Any})

Returns the results from the specified query. Only the fields requested in the query are returned, along with a @ptr field, which is the identifier for the log record. You can use the value of @ptr in a GetLogRecord operation to get the full log record. GetQueryResults does not start a query execution. To run a query, use StartQuery. If the value of the Status field in the output is Running, this operation returns only partial results. If you see a value of Scheduled or Running for the status, you can retry the operation later to see the final results.

Arguments

query_id: The ID number of the query.

Main.Cloudwatch_Logs.list_tags_log_group — Method

list_tags_log_group(log_group_name)
list_tags_log_group(log_group_name, params::Dict{String,<:Any})

Lists the tags for the specified log group.

Arguments

log_group_name: The name of the log group.

Main.Cloudwatch_Logs.put_destination — Method

put_destination(destination_name, role_arn, target_arn)
put_destination(destination_name, role_arn, target_arn, params::Dict{String,<:Any})

Creates or updates a destination. This operation is used only to create destinations for cross-account subscriptions. A destination encapsulates a physical resource (such as an Amazon Kinesis stream) and enables you to subscribe to a real-time stream of log events for a different account, ingested using PutLogEvents. Through an access policy, a destination controls what is written to it. By default, PutDestination does not set any access policy with the destination, which means a cross-account user cannot call PutSubscriptionFilter against this destination. To enable this, the destination owner must call PutDestinationPolicy after PutDestination. To perform a PutDestination operation, you must also have the iam:PassRole permission.

Arguments

destination_name: A name for the destination.

role_arn: The ARN of an IAM role that grants CloudWatch Logs permissions to call the Amazon Kinesis PutRecord operation on the destination stream.

target_arn: The ARN of an Amazon Kinesis stream to which to deliver matching log events.

Main.Cloudwatch_Logs.put_destination_policy — Method

put_destination_policy(access_policy, destination_name)
put_destination_policy(access_policy, destination_name, params::Dict{String,<:Any})

Creates or updates an access policy associated with an existing destination. An access policy is an IAM policy document that is used to authorize claims to register a subscription filter against a given destination.

Arguments

access_policy: An IAM policy document that authorizes cross-account users to deliver their log events to the associated destination. This can be up to 5120 bytes.

destination_name: A name for an existing destination.

Main.Cloudwatch_Logs.put_log_events — Method

put_log_events(log_events, log_group_name, log_stream_name)
put_log_events(log_events, log_group_name, log_stream_name, params::Dict{String,<:Any})

Uploads a batch of log events to the specified log stream. You must include the sequence token obtained from the response of the previous call. An upload in a newly created log stream does not require a sequence token. You can also get the sequence token in the expectedSequenceToken field from InvalidSequenceTokenException. If you call PutLogEvents twice within a narrow time period using the same value for sequenceToken, both calls might be successful or one might be rejected. The batch of events must satisfy the following constraints: The maximum batch size is 1,048,576 bytes. This size is calculated as the sum of all event messages in UTF-8, plus 26 bytes for each log event. None of the log events in the batch can be more than 2 hours in the future. None of the log events in the batch can be older than 14 days or older than the retention period of the log group. The log events in the batch must be in chronological order by their timestamp. The timestamp is the time the event occurred, expressed as the number of milliseconds after Jan 1, 1970 00:00:00 UTC. (In AWS Tools for PowerShell and the AWS SDK for .NET, the timestamp is specified in .NET format: yyyy-mm-ddThh:mm:ss. For example, 2017-09-15T13:45:30.) A batch of log events in a single request cannot span more than 24 hours. Otherwise, the operation fails. The maximum number of log events in a batch is 10,000. There is a quota of 5 requests per second per log stream. Additional requests are throttled. This quota can't be changed. If a call to PutLogEvents returns "UnrecognizedClientException" the most likely cause is an invalid AWS access key ID or secret key.

Arguments

log_events: The log events.

log_group_name: The name of the log group.

log_stream_name: The name of the log stream.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"sequenceToken": The sequence token obtained from the response of the previous PutLogEvents call. An upload in a newly created log stream does not require a sequence token. You can also get the sequence token using DescribeLogStreams. If you call PutLogEvents twice within a narrow time period using the same value for sequenceToken, both calls might be successful or one might be rejected.

Main.Cloudwatch_Logs.put_metric_filter — Method

put_metric_filter(filter_name, filter_pattern, log_group_name, metric_transformations)
put_metric_filter(filter_name, filter_pattern, log_group_name, metric_transformations, params::Dict{String,<:Any})

Creates or updates a metric filter and associates it with the specified log group. Metric filters allow you to configure rules to extract metric data from log events ingested through PutLogEvents. The maximum number of metric filters that can be associated with a log group is 100.

Arguments

filter_name: A name for the metric filter.

filter_pattern: A filter pattern for extracting metric data out of ingested log events.

log_group_name: The name of the log group.

metric_transformations: A collection of information that defines how metric data gets emitted.

Main.Cloudwatch_Logs.put_query_definition — Method

put_query_definition(name, query_string)
put_query_definition(name, query_string, params::Dict{String,<:Any})

Creates or updates a query definition for CloudWatch Logs Insights. For more information, see Analyzing Log Data with CloudWatch Logs Insights. To update a query definition, specify its queryDefinitionId in your request. The values of name, queryString, and logGroupNames are changed to the values that you specify in your update operation. No current values are retained from the current query definition. For example, if you update a current query definition that includes log groups, and you don't specify the logGroupNames parameter in your update operation, the query definition changes to contain no log groups. You must have the logs:PutQueryDefinition permission to be able to perform this operation.

Arguments

name: A name for the query definition. If you are saving a lot of query definitions, we recommend that you name them so that you can easily find the ones you want by using the first part of the name as a filter in the queryDefinitionNamePrefix parameter of DescribeQueryDefinitions.

query_string: The query string to use for this definition. For more information, see CloudWatch Logs Insights Query Syntax.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"logGroupNames": Use this parameter to include specific log groups as part of your query definition. If you are updating a query definition and you omit this parameter, then the updated definition will contain no log groups.

"queryDefinitionId": If you are updating a query definition, use this parameter to specify the ID of the query definition that you want to update. You can use DescribeQueryDefinitions to retrieve the IDs of your saved query definitions. If you are creating a query definition, do not specify this parameter. CloudWatch generates a unique ID for the new query definition and include it in the response to this operation.

Main.Cloudwatch_Logs.put_resource_policy — Method

put_resource_policy()
put_resource_policy(params::Dict{String,<:Any})

Creates or updates a resource policy allowing other AWS services to put log events to this account, such as Amazon Route 53. An account can have up to 10 resource policies per AWS Region.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"policyDocument": Details of the new policy, including the identity of the principal that is enabled to put logs to this account. This is formatted as a JSON string. This parameter is required. The following example creates a resource policy enabling the Route 53 service to put DNS query logs in to the specified log group. Replace "logArn" with the ARN of your CloudWatch Logs resource, such as a log group or log stream. { "Version": "2012-10-17", "Statement": [ { "Sid": "Route53LogsToCloudWatchLogs", "Effect": "Allow", "Principal": { "Service": [ "route53.amazonaws.com" ] }, "Action":"logs:PutLogEvents", "Resource": "logArn" } ] }

"policyName": Name of the new policy. This parameter is required.

Main.Cloudwatch_Logs.put_retention_policy — Method

put_retention_policy(log_group_name, retention_in_days)
put_retention_policy(log_group_name, retention_in_days, params::Dict{String,<:Any})

Sets the retention of the specified log group. A retention policy allows you to configure the number of days for which to retain log events in the specified log group.

Arguments

log_group_name: The name of the log group.

retention_in_days:

Main.Cloudwatch_Logs.put_subscription_filter — Method

put_subscription_filter(destination_arn, filter_name, filter_pattern, log_group_name)
put_subscription_filter(destination_arn, filter_name, filter_pattern, log_group_name, params::Dict{String,<:Any})

Creates or updates a subscription filter and associates it with the specified log group. Subscription filters allow you to subscribe to a real-time stream of log events ingested through PutLogEvents and have them delivered to a specific destination. When log events are sent to the receiving service, they are Base64 encoded and compressed with the gzip format. The following destinations are supported for subscription filters: An Amazon Kinesis stream belonging to the same account as the subscription filter, for same-account delivery. A logical destination that belongs to a different account, for cross-account delivery. An Amazon Kinesis Firehose delivery stream that belongs to the same account as the subscription filter, for same-account delivery. An AWS Lambda function that belongs to the same account as the subscription filter, for same-account delivery. There can only be one subscription filter associated with a log group. If you are updating an existing filter, you must specify the correct name in filterName. Otherwise, the call fails because you cannot associate a second filter with a log group. To perform a PutSubscriptionFilter operation, you must also have the iam:PassRole permission.

Arguments

destination_arn: The ARN of the destination to deliver matching log events to. Currently, the supported destinations are: An Amazon Kinesis stream belonging to the same account as the subscription filter, for same-account delivery. A logical destination (specified using an ARN) belonging to a different account, for cross-account delivery. An Amazon Kinesis Firehose delivery stream belonging to the same account as the subscription filter, for same-account delivery. An AWS Lambda function belonging to the same account as the subscription filter, for same-account delivery.

filter_name: A name for the subscription filter. If you are updating an existing filter, you must specify the correct name in filterName. Otherwise, the call fails because you cannot associate a second filter with a log group. To find the name of the filter currently associated with a log group, use DescribeSubscriptionFilters.

filter_pattern: A filter pattern for subscribing to a filtered stream of log events.

log_group_name: The name of the log group.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"distribution": The method used to distribute log data to the destination. By default, log data is grouped by log stream, but the grouping can be set to random for a more even distribution. This property is only applicable when the destination is an Amazon Kinesis stream.

"roleArn": The ARN of an IAM role that grants CloudWatch Logs permissions to deliver ingested log events to the destination stream. You don't need to provide the ARN when you are working with a logical destination for cross-account delivery.

Main.Cloudwatch_Logs.start_query — Method

start_query(end_time, query_string, start_time)
start_query(end_time, query_string, start_time, params::Dict{String,<:Any})

Schedules a query of a log group using CloudWatch Logs Insights. You specify the log group and time range to query and the query string to use. For more information, see CloudWatch Logs Insights Query Syntax. Queries time out after 15 minutes of execution. If your queries are timing out, reduce the time range being searched or partition your query into a number of queries.

Arguments

end_time: The end of the time range to query. The range is inclusive, so the specified end time is included in the query. Specified as epoch time, the number of seconds since January 1, 1970, 00:00:00 UTC.

query_string: The query string to use. For more information, see CloudWatch Logs Insights Query Syntax.

start_time: The beginning of the time range to query. The range is inclusive, so the specified start time is included in the query. Specified as epoch time, the number of seconds since January 1, 1970, 00:00:00 UTC.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"limit": The maximum number of log events to return in the query. If the query string uses the fields command, only the specified fields and their values are returned. The default is 1000.

"logGroupName": The log group on which to perform the query. A StartQuery operation must include a logGroupNames or a logGroupName parameter, but not both.

"logGroupNames": The list of log groups to be queried. You can include up to 20 log groups. A StartQuery operation must include a logGroupNames or a logGroupName parameter, but not both.

Main.Cloudwatch_Logs.stop_query — Method

stop_query(query_id)
stop_query(query_id, params::Dict{String,<:Any})

Stops a CloudWatch Logs Insights query that is in progress. If the query has already ended, the operation returns an error indicating that the specified query is not running.

Arguments

query_id: The ID number of the query to stop. To find this ID number, use DescribeQueries.

Main.Cloudwatch_Logs.tag_log_group — Method

tag_log_group(log_group_name, tags)
tag_log_group(log_group_name, tags, params::Dict{String,<:Any})

Adds or updates the specified tags for the specified log group. To list the tags for a log group, use ListTagsLogGroup. To remove tags, use UntagLogGroup. For more information about tags, see Tag Log Groups in Amazon CloudWatch Logs in the Amazon CloudWatch Logs User Guide.

Arguments

log_group_name: The name of the log group.

tags: The key-value pairs to use for the tags.

Main.Cloudwatch_Logs.test_metric_filter — Method

test_metric_filter(filter_pattern, log_event_messages)
test_metric_filter(filter_pattern, log_event_messages, params::Dict{String,<:Any})

Tests the filter pattern of a metric filter against a sample of log event messages. You can use this operation to validate the correctness of a metric filter pattern.

Arguments

filter_pattern:

log_event_messages: The log event messages to test.

Main.Cloudwatch_Logs.untag_log_group — Method

untag_log_group(log_group_name, tags)
untag_log_group(log_group_name, tags, params::Dict{String,<:Any})

Removes the specified tags from the specified log group. To list the tags for a log group, use ListTagsLogGroup. To add tags, use TagLogGroup.

Arguments

log_group_name: The name of the log group.

tags: The tag keys. The corresponding tags are removed from the log group.