Codeguru Reviewer

This page documents function available when using the Codeguru_Reviewer module, created with @service Codeguru_Reviewer.

Index

Main.Codeguru_Reviewer.associate_repository

Main.Codeguru_Reviewer.create_code_review

Main.Codeguru_Reviewer.describe_code_review

Main.Codeguru_Reviewer.describe_recommendation_feedback

Main.Codeguru_Reviewer.describe_repository_association

Main.Codeguru_Reviewer.disassociate_repository

Main.Codeguru_Reviewer.list_code_reviews

Main.Codeguru_Reviewer.list_recommendation_feedback

Main.Codeguru_Reviewer.list_recommendations

Main.Codeguru_Reviewer.list_repository_associations

Main.Codeguru_Reviewer.list_tags_for_resource

Main.Codeguru_Reviewer.put_recommendation_feedback

Main.Codeguru_Reviewer.tag_resource

Main.Codeguru_Reviewer.untag_resource

Documentation

Main.Codeguru_Reviewer.associate_repository — Method

associate_repository(repository)
associate_repository(repository, params::Dict{String,<:Any})

Use to associate an Amazon Web Services CodeCommit repository or a repostory managed by Amazon Web Services CodeStar Connections with Amazon CodeGuru Reviewer. When you associate a repository, CodeGuru Reviewer reviews source code changes in the repository's pull requests and provides automatic recommendations. You can view recommendations using the CodeGuru Reviewer console. For more information, see Recommendations in Amazon CodeGuru Reviewer in the Amazon CodeGuru Reviewer User Guide. If you associate a CodeCommit or S3 repository, it must be in the same Amazon Web Services Region and Amazon Web Services account where its CodeGuru Reviewer code reviews are configured. Bitbucket and GitHub Enterprise Server repositories are managed by Amazon Web Services CodeStar Connections to connect to CodeGuru Reviewer. For more information, see Associate a repository in the Amazon CodeGuru Reviewer User Guide. You cannot use the CodeGuru Reviewer SDK or the Amazon Web Services CLI to associate a GitHub repository with Amazon CodeGuru Reviewer. To associate a GitHub repository, use the console. For more information, see Getting started with CodeGuru Reviewer in the CodeGuru Reviewer User Guide.

Arguments

repository: The repository to associate.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientRequestToken": Amazon CodeGuru Reviewer uses this value to prevent the accidental creation of duplicate repository associations if there are failures and retries.

"KMSKeyDetails": A KMSKeyDetails object that contains: The encryption option for this repository association. It is either owned by Amazon Web Services Key Management Service (KMS) (AWSOWNEDCMK) or customer managed (CUSTOMERMANAGEDCMK). The ID of the Amazon Web Services KMS key that is associated with this respository association.

"Tags": An array of key-value pairs used to tag an associated repository. A tag is a custom attribute label with two parts: A tag key (for example, CostCenter, Environment, Project, or Secret). Tag keys are case sensitive. An optional field known as a tag value (for example, 111122223333, Production, or a team name). Omitting the tag value is the same as using an empty string. Like tag keys, tag values are case sensitive.

Main.Codeguru_Reviewer.create_code_review — Method

create_code_review(name, repository_association_arn, type)
create_code_review(name, repository_association_arn, type, params::Dict{String,<:Any})

Use to create a code review with a CodeReviewType of RepositoryAnalysis. This type of code review analyzes all code under a specified branch in an associated repository. PullRequest code reviews are automatically triggered by a pull request.

Arguments

name: The name of the code review. The name of each code review in your Amazon Web Services account must be unique.

repository_association_arn: The Amazon Resource Name (ARN) of the RepositoryAssociation object. You can retrieve this ARN by calling ListRepositoryAssociations . A code review can only be created on an associated repository. This is the ARN of the associated repository.

type: The type of code review to create. This is specified using a CodeReviewType object. You can create a code review only of type RepositoryAnalysis.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientRequestToken": Amazon CodeGuru Reviewer uses this value to prevent the accidental creation of duplicate code reviews if there are failures and retries.

Main.Codeguru_Reviewer.describe_code_review — Method

describe_code_review(code_review_arn)
describe_code_review(code_review_arn, params::Dict{String,<:Any})

Returns the metadata associated with the code review along with its status.

Arguments

code_review_arn: The Amazon Resource Name (ARN) of the CodeReview object.

Main.Codeguru_Reviewer.describe_recommendation_feedback — Method

describe_recommendation_feedback(code_review_arn, recommendation_id)
describe_recommendation_feedback(code_review_arn, recommendation_id, params::Dict{String,<:Any})

Describes the customer feedback for a CodeGuru Reviewer recommendation.

Arguments

code_review_arn: The Amazon Resource Name (ARN) of the CodeReview object.

recommendation_id: The recommendation ID that can be used to track the provided recommendations and then to collect the feedback.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"UserId": Optional parameter to describe the feedback for a given user. If this is not supplied, it defaults to the user making the request. The UserId is an IAM principal that can be specified as an Amazon Web Services account ID or an Amazon Resource Name (ARN). For more information, see Specifying a Principal in the Amazon Web Services Identity and Access Management User Guide.

Main.Codeguru_Reviewer.describe_repository_association — Method

describe_repository_association(association_arn)
describe_repository_association(association_arn, params::Dict{String,<:Any})

Returns a RepositoryAssociation object that contains information about the requested repository association.

Arguments

association_arn: The Amazon Resource Name (ARN) of the RepositoryAssociation object. You can retrieve this ARN by calling ListRepositoryAssociations .

Main.Codeguru_Reviewer.disassociate_repository — Method

disassociate_repository(association_arn)
disassociate_repository(association_arn, params::Dict{String,<:Any})

Removes the association between Amazon CodeGuru Reviewer and a repository.

Arguments

association_arn: The Amazon Resource Name (ARN) of the RepositoryAssociation object. You can retrieve this ARN by calling ListRepositoryAssociations .

Main.Codeguru_Reviewer.list_code_reviews — Method

list_code_reviews(type)
list_code_reviews(type, params::Dict{String,<:Any})

Lists all the code reviews that the customer has created in the past 90 days.

Arguments

type: The type of code reviews to list in the response.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results that are returned per call. The default is

"NextToken": If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

"ProviderTypes": List of provider types for filtering that needs to be applied before displaying the result. For example, providerTypes=[GitHub] lists code reviews from GitHub.

"RepositoryNames": List of repository names for filtering that needs to be applied before displaying the result.

"States": List of states for filtering that needs to be applied before displaying the result. For example, states=[Pending] lists code reviews in the Pending state. The valid code review states are: Completed: The code review is complete. Pending: The code review started and has not completed or failed. Failed: The code review failed. Deleting: The code review is being deleted.

Main.Codeguru_Reviewer.list_recommendation_feedback — Method

list_recommendation_feedback(code_review_arn)
list_recommendation_feedback(code_review_arn, params::Dict{String,<:Any})

Returns a list of RecommendationFeedbackSummary objects that contain customer recommendation feedback for all CodeGuru Reviewer users.

Arguments

code_review_arn: The Amazon Resource Name (ARN) of the CodeReview object.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results that are returned per call. The default is

"NextToken": If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged.

"RecommendationIds": Used to query the recommendation feedback for a given recommendation.

"UserIds": An Amazon Web Services user's account ID or Amazon Resource Name (ARN). Use this ID to query the recommendation feedback for a code review from that user. The UserId is an IAM principal that can be specified as an Amazon Web Services account ID or an Amazon Resource Name (ARN). For more information, see Specifying a Principal in the Amazon Web Services Identity and Access Management User Guide.

Main.Codeguru_Reviewer.list_recommendations — Method

list_recommendations(code_review_arn)
list_recommendations(code_review_arn, params::Dict{String,<:Any})

Returns the list of all recommendations for a completed code review.

Arguments

code_review_arn: The Amazon Resource Name (ARN) of the CodeReview object.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results that are returned per call. The default is

"NextToken": Pagination token.

Main.Codeguru_Reviewer.list_repository_associations — Method

list_repository_associations()
list_repository_associations(params::Dict{String,<:Any})

Returns a list of RepositoryAssociationSummary objects that contain summary information about a repository association. You can filter the returned list by ProviderType , Name , State , and Owner .

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of repository association results returned by ListRepositoryAssociations in paginated output. When this parameter is used, ListRepositoryAssociations only returns maxResults results in a single page with a nextToken response element. The remaining results of the initial request can be seen by sending another ListRepositoryAssociations request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, ListRepositoryAssociations returns up to 100 results and a nextToken value if applicable.

"Name": List of repository names to use as a filter.

"NextToken": The nextToken value returned from a previous paginated ListRepositoryAssociations request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. Treat this token as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

"Owner": List of owners to use as a filter. For Amazon Web Services CodeCommit, it is the name of the CodeCommit account that was used to associate the repository. For other repository source providers, such as Bitbucket and GitHub Enterprise Server, this is name of the account that was used to associate the repository.

"ProviderType": List of provider types to use as a filter.

"State": List of repository association states to use as a filter. The valid repository association states are: Associated: The repository association is complete. Associating: CodeGuru Reviewer is: Setting up pull request notifications. This is required for pull requests to trigger a CodeGuru Reviewer review. If your repository ProviderType is GitHub, GitHub Enterprise Server, or Bitbucket, CodeGuru Reviewer creates webhooks in your repository to trigger CodeGuru Reviewer reviews. If you delete these webhooks, reviews of code in your repository cannot be triggered. Setting up source code access. This is required for CodeGuru Reviewer to securely clone code in your repository. Failed: The repository failed to associate or disassociate. Disassociating: CodeGuru Reviewer is removing the repository's pull request notifications and source code access. Disassociated: CodeGuru Reviewer successfully disassociated the repository. You can create a new association with this repository if you want to review source code in it later. You can control access to code reviews created in an associated repository with tags after it has been disassociated. For more information, see Using tags to control access to associated repositories in the Amazon CodeGuru Reviewer User Guide.

Main.Codeguru_Reviewer.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Returns the list of tags associated with an associated repository resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the RepositoryAssociation object. You can retrieve this ARN by calling ListRepositoryAssociations .

Main.Codeguru_Reviewer.put_recommendation_feedback — Method

put_recommendation_feedback(code_review_arn, reactions, recommendation_id)
put_recommendation_feedback(code_review_arn, reactions, recommendation_id, params::Dict{String,<:Any})

Stores customer feedback for a CodeGuru Reviewer recommendation. When this API is called again with different reactions the previous feedback is overwritten.

Arguments

code_review_arn: The Amazon Resource Name (ARN) of the CodeReview object.

reactions: List for storing reactions. Reactions are utf-8 text code for emojis. If you send an empty list it clears all your feedback.

recommendation_id: The recommendation ID that can be used to track the provided recommendations and then to collect the feedback.

Main.Codeguru_Reviewer.tag_resource — Method

tag_resource(tags, resource_arn)
tag_resource(tags, resource_arn, params::Dict{String,<:Any})

Adds one or more tags to an associated repository.

Arguments

tags: An array of key-value pairs used to tag an associated repository. A tag is a custom attribute label with two parts: A tag key (for example, CostCenter, Environment, Project, or Secret). Tag keys are case sensitive. An optional field known as a tag value (for example, 111122223333, Production, or a team name). Omitting the tag value is the same as using an empty string. Like tag keys, tag values are case sensitive.

resource_arn: The Amazon Resource Name (ARN) of the RepositoryAssociation object. You can retrieve this ARN by calling ListRepositoryAssociations .

Main.Codeguru_Reviewer.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes a tag from an associated repository.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the RepositoryAssociation object. You can retrieve this ARN by calling ListRepositoryAssociations .

tag_keys: A list of the keys for each tag you want to remove from an associated repository.