Codestar Notifications

This page documents function available when using the Codestar_Notifications module, created with @service Codestar_Notifications.

Index

Main.Codestar_Notifications.create_notification_rule

Main.Codestar_Notifications.delete_notification_rule

Main.Codestar_Notifications.delete_target

Main.Codestar_Notifications.describe_notification_rule

Main.Codestar_Notifications.list_event_types

Main.Codestar_Notifications.list_notification_rules

Main.Codestar_Notifications.list_tags_for_resource

Main.Codestar_Notifications.list_targets

Main.Codestar_Notifications.subscribe

Main.Codestar_Notifications.tag_resource

Main.Codestar_Notifications.unsubscribe

Main.Codestar_Notifications.untag_resource

Main.Codestar_Notifications.update_notification_rule

Documentation

Main.Codestar_Notifications.create_notification_rule — Method

create_notification_rule(detail_type, event_type_ids, name, resource, targets)
create_notification_rule(detail_type, event_type_ids, name, resource, targets, params::Dict{String,<:Any})

Creates a notification rule for a resource. The rule specifies the events you want notifications about and the targets (such as SNS topics) where you want to receive them.

Arguments

detail_type: The level of detail to include in the notifications for this resource. BASIC will include only the contents of the event as it would appear in AWS CloudWatch. FULL will include any supplemental information provided by AWS CodeStar Notifications and/or the service for the resource for which the notification is created.

event_type_ids: A list of event types associated with this notification rule. For a list of allowed events, see EventTypeSummary.

name: The name for the notification rule. Notifictaion rule names must be unique in your AWS account.

resource: The Amazon Resource Name (ARN) of the resource to associate with the notification rule. Supported resources include pipelines in AWS CodePipeline, repositories in AWS CodeCommit, and build projects in AWS CodeBuild.

targets: A list of Amazon Resource Names (ARNs) of SNS topics to associate with the notification rule.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientRequestToken": A unique, client-generated idempotency token that, when provided in a request, ensures the request cannot be repeated with a changed parameter. If a request with the same parameters is received and a token is included, the request returns information about the initial request that used that token. The AWS SDKs prepopulate client request tokens. If you are using an AWS SDK, an idempotency token is created for you.

"Status": The status of the notification rule. The default value is ENABLED. If the status is set to DISABLED, notifications aren't sent for the notification rule.

"Tags": A list of tags to apply to this notification rule. Key names cannot start with "aws".

Main.Codestar_Notifications.delete_notification_rule — Method

delete_notification_rule(arn)
delete_notification_rule(arn, params::Dict{String,<:Any})

Deletes a notification rule for a resource.

Arguments

arn: The Amazon Resource Name (ARN) of the notification rule you want to delete.

Main.Codestar_Notifications.delete_target — Method

delete_target(target_address)
delete_target(target_address, params::Dict{String,<:Any})

Deletes a specified target for notifications.

Arguments

target_address: The Amazon Resource Name (ARN) of the SNS topic to delete.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ForceUnsubscribeAll": A Boolean value that can be used to delete all associations with this SNS topic. The default value is FALSE. If set to TRUE, all associations between that target and every notification rule in your AWS account are deleted.

Main.Codestar_Notifications.describe_notification_rule — Method

describe_notification_rule(arn)
describe_notification_rule(arn, params::Dict{String,<:Any})

Returns information about a specified notification rule.

Arguments

arn: The Amazon Resource Name (ARN) of the notification rule.

Main.Codestar_Notifications.list_event_types — Method

list_event_types()
list_event_types(params::Dict{String,<:Any})

Returns information about the event types available for configuring notifications.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Filters": The filters to use to return information by service or resource type.

"MaxResults": A non-negative integer used to limit the number of returned results. The default number is 50. The maximum number of results that can be returned is 100.

"NextToken": An enumeration token that, when provided in a request, returns the next batch of the results.

Main.Codestar_Notifications.list_notification_rules — Method

list_notification_rules()
list_notification_rules(params::Dict{String,<:Any})

Returns a list of the notification rules for an AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Filters": The filters to use to return information by service or resource type. For valid values, see ListNotificationRulesFilter. A filter with the same name can appear more than once when used with OR statements. Filters with different names should be applied with AND statements.

"MaxResults": A non-negative integer used to limit the number of returned results. The maximum number of results that can be returned is 100.

"NextToken": An enumeration token that, when provided in a request, returns the next batch of the results.

Main.Codestar_Notifications.list_tags_for_resource — Method

list_tags_for_resource(arn)
list_tags_for_resource(arn, params::Dict{String,<:Any})

Returns a list of the tags associated with a notification rule.

Arguments

arn: The Amazon Resource Name (ARN) for the notification rule.

Main.Codestar_Notifications.list_targets — Method

list_targets()
list_targets(params::Dict{String,<:Any})

Returns a list of the notification rule targets for an AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Filters": The filters to use to return information by service or resource type. Valid filters include target type, target address, and target status. A filter with the same name can appear more than once when used with OR statements. Filters with different names should be applied with AND statements.

"MaxResults": A non-negative integer used to limit the number of returned results. The maximum number of results that can be returned is 100.

"NextToken": An enumeration token that, when provided in a request, returns the next batch of the results.

Main.Codestar_Notifications.subscribe — Method

subscribe(arn, target)
subscribe(arn, target, params::Dict{String,<:Any})

Creates an association between a notification rule and an SNS topic so that the associated target can receive notifications when the events described in the rule are triggered.

Arguments

arn: The Amazon Resource Name (ARN) of the notification rule for which you want to create the association.

target:

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientRequestToken": An enumeration token that, when provided in a request, returns the next batch of the results.

Main.Codestar_Notifications.tag_resource — Method

tag_resource(arn, tags)
tag_resource(arn, tags, params::Dict{String,<:Any})

Associates a set of provided tags with a notification rule.

Arguments

arn: The Amazon Resource Name (ARN) of the notification rule to tag.

tags: The list of tags to associate with the resource. Tag key names cannot start with "aws".

Main.Codestar_Notifications.unsubscribe — Method

unsubscribe(arn, target_address)
unsubscribe(arn, target_address, params::Dict{String,<:Any})

Removes an association between a notification rule and an Amazon SNS topic so that subscribers to that topic stop receiving notifications when the events described in the rule are triggered.

Arguments

arn: The Amazon Resource Name (ARN) of the notification rule.

target_address: The ARN of the SNS topic to unsubscribe from the notification rule.

Main.Codestar_Notifications.untag_resource — Method

untag_resource(arn, tag_keys)
untag_resource(arn, tag_keys, params::Dict{String,<:Any})

Removes the association between one or more provided tags and a notification rule.

Arguments

arn: The Amazon Resource Name (ARN) of the notification rule from which to remove the tags.

tag_keys: The key names of the tags to remove.

Main.Codestar_Notifications.update_notification_rule — Method

update_notification_rule(arn)
update_notification_rule(arn, params::Dict{String,<:Any})

Updates a notification rule for a resource. You can change the events that trigger the notification rule, the status of the rule, and the targets that receive the notifications. To add or remove tags for a notification rule, you must use TagResource and UntagResource.

Arguments

arn: The Amazon Resource Name (ARN) of the notification rule.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DetailType": The level of detail to include in the notifications for this resource. BASIC will include only the contents of the event as it would appear in AWS CloudWatch. FULL will include any supplemental information provided by AWS CodeStar Notifications and/or the service for the resource for which the notification is created.

"EventTypeIds": A list of event types associated with this notification rule.

"Name": The name of the notification rule.

"Status": The status of the notification rule. Valid statuses include enabled (sending notifications) or disabled (not sending notifications).

"Targets": The address and type of the targets to receive notifications from this notification rule.