Device Farm

This page documents function available when using the Device_Farm module, created with @service Device_Farm.

Index

Main.Device_Farm.create_device_pool

Main.Device_Farm.create_instance_profile

Main.Device_Farm.create_network_profile

Main.Device_Farm.create_project

Main.Device_Farm.create_remote_access_session

Main.Device_Farm.create_test_grid_project

Main.Device_Farm.create_test_grid_url

Main.Device_Farm.create_upload

Main.Device_Farm.create_vpceconfiguration

Main.Device_Farm.delete_device_pool

Main.Device_Farm.delete_instance_profile

Main.Device_Farm.delete_network_profile

Main.Device_Farm.delete_project

Main.Device_Farm.delete_remote_access_session

Main.Device_Farm.delete_run

Main.Device_Farm.delete_test_grid_project

Main.Device_Farm.delete_upload

Main.Device_Farm.delete_vpceconfiguration

Main.Device_Farm.get_account_settings

Main.Device_Farm.get_device

Main.Device_Farm.get_device_instance

Main.Device_Farm.get_device_pool

Main.Device_Farm.get_device_pool_compatibility

Main.Device_Farm.get_instance_profile

Main.Device_Farm.get_job

Main.Device_Farm.get_network_profile

Main.Device_Farm.get_offering_status

Main.Device_Farm.get_project

Main.Device_Farm.get_remote_access_session

Main.Device_Farm.get_run

Main.Device_Farm.get_suite

Main.Device_Farm.get_test

Main.Device_Farm.get_test_grid_project

Main.Device_Farm.get_test_grid_session

Main.Device_Farm.get_upload

Main.Device_Farm.get_vpceconfiguration

Main.Device_Farm.install_to_remote_access_session

Main.Device_Farm.list_artifacts

Main.Device_Farm.list_device_instances

Main.Device_Farm.list_device_pools

Main.Device_Farm.list_devices

Main.Device_Farm.list_instance_profiles

Main.Device_Farm.list_jobs

Main.Device_Farm.list_network_profiles

Main.Device_Farm.list_offering_promotions

Main.Device_Farm.list_offering_transactions

Main.Device_Farm.list_offerings

Main.Device_Farm.list_projects

Main.Device_Farm.list_remote_access_sessions

Main.Device_Farm.list_runs

Main.Device_Farm.list_samples

Main.Device_Farm.list_suites

Main.Device_Farm.list_tags_for_resource

Main.Device_Farm.list_test_grid_projects

Main.Device_Farm.list_test_grid_session_actions

Main.Device_Farm.list_test_grid_session_artifacts

Main.Device_Farm.list_test_grid_sessions

Main.Device_Farm.list_tests

Main.Device_Farm.list_unique_problems

Main.Device_Farm.list_uploads

Main.Device_Farm.list_vpceconfigurations

Main.Device_Farm.purchase_offering

Main.Device_Farm.renew_offering

Main.Device_Farm.schedule_run

Main.Device_Farm.stop_job

Main.Device_Farm.stop_remote_access_session

Main.Device_Farm.stop_run

Main.Device_Farm.tag_resource

Main.Device_Farm.untag_resource

Main.Device_Farm.update_device_instance

Main.Device_Farm.update_device_pool

Main.Device_Farm.update_instance_profile

Main.Device_Farm.update_network_profile

Main.Device_Farm.update_project

Main.Device_Farm.update_test_grid_project

Main.Device_Farm.update_upload

Main.Device_Farm.update_vpceconfiguration

Documentation

Main.Device_Farm.create_device_pool — Method

create_device_pool(name, project_arn, rules)
create_device_pool(name, project_arn, rules, params::Dict{String,<:Any})

Creates a device pool.

Arguments

name: The device pool's name.

project_arn: The ARN of the project for the device pool.

rules: The device pool's rules.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The device pool's description.

"maxDevices": The number of devices that Device Farm can add to your device pool. Device Farm adds devices that are available and meet the criteria that you assign for the rules parameter. Depending on how many devices meet these constraints, your device pool might contain fewer devices than the value for this parameter. By specifying the maximum number of devices, you can control the costs that you incur by running tests.

Main.Device_Farm.create_instance_profile — Method

create_instance_profile(name)
create_instance_profile(name, params::Dict{String,<:Any})

Creates a profile that can be applied to one or more private fleet device instances.

Arguments

name: The name of your instance profile.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The description of your instance profile.

"excludeAppPackagesFromCleanup": An array of strings that specifies the list of app packages that should not be cleaned up from the device after a test run. The list of packages is considered only if you set packageCleanup to true.

"packageCleanup": When set to true, Device Farm removes app packages after a test run. The default value is false for private devices.

"rebootAfterUse": When set to true, Device Farm reboots the instance after a test run. The default value is true.

Main.Device_Farm.create_network_profile — Method

create_network_profile(name, project_arn)
create_network_profile(name, project_arn, params::Dict{String,<:Any})

Creates a network profile.

Arguments

name: The name for the new network profile.

project_arn: The Amazon Resource Name (ARN) of the project for which you want to create a network profile.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The description of the network profile.

"downlinkBandwidthBits": The data throughput rate in bits per second, as an integer from 0 to 104857600.

"downlinkDelayMs": Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

"downlinkJitterMs": Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

"downlinkLossPercent": Proportion of received packets that fail to arrive from 0 to 100 percent.

"type": The type of network profile to create. Valid values are listed here.

"uplinkBandwidthBits": The data throughput rate in bits per second, as an integer from 0 to 104857600.

"uplinkDelayMs": Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

"uplinkJitterMs": Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

"uplinkLossPercent": Proportion of transmitted packets that fail to arrive from 0 to 100 percent.

Main.Device_Farm.create_project — Method

create_project(name)
create_project(name, params::Dict{String,<:Any})

Creates a project.

Arguments

name: The project's name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"defaultJobTimeoutMinutes": Sets the execution timeout value (in minutes) for a project. All test runs in this project use the specified execution timeout value unless overridden when scheduling a run.

"vpcConfig": The VPC security groups and subnets that are attached to a project.

Main.Device_Farm.create_remote_access_session — Method

create_remote_access_session(device_arn, project_arn)
create_remote_access_session(device_arn, project_arn, params::Dict{String,<:Any})

Specifies and starts a remote access session.

Arguments

device_arn: The ARN of the device for which you want to create a remote access session.

project_arn: The Amazon Resource Name (ARN) of the project for which you want to create a remote access session.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientId": Unique identifier for the client. If you want access to multiple devices on the same client, you should pass the same clientId value in each call to CreateRemoteAccessSession. This identifier is required only if remoteDebugEnabled is set to true. Remote debugging is no longer supported.

"configuration": The configuration information for the remote access session request.

"instanceArn": The Amazon Resource Name (ARN) of the device instance for which you want to create a remote access session.

"interactionMode": The interaction mode of the remote access session. Valid values are: INTERACTIVE: You can interact with the iOS device by viewing, touching, and rotating the screen. You cannot run XCUITest framework-based tests in this mode. NOVIDEO: You are connected to the device, but cannot interact with it or view the screen. This mode has the fastest test execution speed. You can run XCUITest framework-based tests in this mode. VIDEOONLY: You can view the screen, but cannot touch or rotate it. You can run XCUITest framework-based tests and watch the screen in this mode.

"name": The name of the remote access session to create.

"remoteDebugEnabled": Set to true if you want to access devices remotely for debugging in your remote access session. Remote debugging is no longer supported.

"remoteRecordAppArn": The Amazon Resource Name (ARN) for the app to be recorded in the remote access session.

"remoteRecordEnabled": Set to true to enable remote recording for the remote access session.

"skipAppResign": When set to true, for private devices, Device Farm does not sign your app again. For public devices, Device Farm always signs your apps again. For more information on how Device Farm modifies your uploads during tests, see Do you modify my app?

"sshPublicKey": Ignored. The public key of the ssh key pair you want to use for connecting to remote devices in your remote debugging session. This key is required only if remoteDebugEnabled is set to true. Remote debugging is no longer supported.

Main.Device_Farm.create_test_grid_project — Method

create_test_grid_project(name)
create_test_grid_project(name, params::Dict{String,<:Any})

Creates a Selenium testing project. Projects are used to track TestGridSession instances.

Arguments

name: Human-readable name of the Selenium testing project.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": Human-readable description of the project.

"vpcConfig": The VPC security groups and subnets that are attached to a project.

Main.Device_Farm.create_test_grid_url — Method

create_test_grid_url(expires_in_seconds, project_arn)
create_test_grid_url(expires_in_seconds, project_arn, params::Dict{String,<:Any})

Creates a signed, short-term URL that can be passed to a Selenium RemoteWebDriver constructor.

Arguments

expires_in_seconds: Lifetime, in seconds, of the URL.

project_arn: ARN (from CreateTestGridProject or ListTestGridProjects) to associate with the short-term URL.

Main.Device_Farm.create_upload — Method

create_upload(name, project_arn, type)
create_upload(name, project_arn, type, params::Dict{String,<:Any})

Uploads an app or test scripts.

Arguments

name: The upload's file name. The name should not contain any forward slashes (/). If you are uploading an iOS app, the file name must end with the .ipa extension. If you are uploading an Android app, the file name must end with the .apk extension. For all others, the file name must end with the .zip file extension.

project_arn: The ARN of the project for the upload.

type: The upload's upload type. Must be one of the following values: ANDROIDAPP IOSAPP WEBAPP EXTERNALDATA APPIUMJAVAJUNITTESTPACKAGE APPIUMJAVATESTNGTESTPACKAGE APPIUMPYTHONTESTPACKAGE APPIUMNODETESTPACKAGE APPIUMRUBYTESTPACKAGE APPIUMWEBJAVAJUNITTESTPACKAGE APPIUMWEBJAVATESTNGTESTPACKAGE APPIUMWEBPYTHONTESTPACKAGE APPIUMWEBNODETESTPACKAGE APPIUMWEBRUBYTESTPACKAGE CALABASHTESTPACKAGE INSTRUMENTATIONTESTPACKAGE UIAUTOMATIONTESTPACKAGE UIAUTOMATORTESTPACKAGE XCTESTTESTPACKAGE XCTESTUITESTPACKAGE APPIUMJAVAJUNITTESTSPEC APPIUMJAVATESTNGTESTSPEC APPIUMPYTHONTESTSPEC APPIUMNODETESTSPEC APPIUMRUBYTESTSPEC APPIUMWEBJAVAJUNITTESTSPEC APPIUMWEBJAVATESTNGTESTSPEC APPIUMWEBPYTHONTESTSPEC APPIUMWEBNODETESTSPEC APPIUMWEBRUBYTESTSPEC INSTRUMENTATIONTESTSPEC XCTESTUITESTSPEC If you call CreateUpload with WEB_APP specified, AWS Device Farm throws an ArgumentException error.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"contentType": The upload's content type (for example, application/octet-stream).

Main.Device_Farm.create_vpceconfiguration — Method

create_vpceconfiguration(service_dns_name, vpce_configuration_name, vpce_service_name)
create_vpceconfiguration(service_dns_name, vpce_configuration_name, vpce_service_name, params::Dict{String,<:Any})

Creates a configuration record in Device Farm for your Amazon Virtual Private Cloud (VPC) endpoint.

Arguments

service_dns_name: The DNS name of the service running in your VPC that you want Device Farm to test.

vpce_configuration_name: The friendly name you give to your VPC endpoint configuration, to manage your configurations more easily.

vpce_service_name: The name of the VPC endpoint service running in your AWS account that you want Device Farm to test.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"vpceConfigurationDescription": An optional description that provides details about your VPC endpoint configuration.

Main.Device_Farm.delete_device_pool — Method

delete_device_pool(arn)
delete_device_pool(arn, params::Dict{String,<:Any})

Deletes a device pool given the pool ARN. Does not allow deletion of curated pools owned by the system.

Arguments

arn: Represents the Amazon Resource Name (ARN) of the Device Farm device pool to delete.

Main.Device_Farm.delete_instance_profile — Method

delete_instance_profile(arn)
delete_instance_profile(arn, params::Dict{String,<:Any})

Deletes a profile that can be applied to one or more private device instances.

Arguments

arn: The Amazon Resource Name (ARN) of the instance profile you are requesting to delete.

Main.Device_Farm.delete_network_profile — Method

delete_network_profile(arn)
delete_network_profile(arn, params::Dict{String,<:Any})

Deletes a network profile.

Arguments

arn: The ARN of the network profile to delete.

Main.Device_Farm.delete_project — Method

delete_project(arn)
delete_project(arn, params::Dict{String,<:Any})

Deletes an AWS Device Farm project, given the project ARN. Deleting this resource does not stop an in-progress run.

Arguments

arn: Represents the Amazon Resource Name (ARN) of the Device Farm project to delete.

Main.Device_Farm.delete_remote_access_session — Method

delete_remote_access_session(arn)
delete_remote_access_session(arn, params::Dict{String,<:Any})

Deletes a completed remote access session and its results.

Arguments

arn: The Amazon Resource Name (ARN) of the session for which you want to delete remote access.

Main.Device_Farm.delete_run — Method

delete_run(arn)
delete_run(arn, params::Dict{String,<:Any})

Deletes the run, given the run ARN. Deleting this resource does not stop an in-progress run.

Arguments

arn: The Amazon Resource Name (ARN) for the run to delete.

Main.Device_Farm.delete_test_grid_project — Method

delete_test_grid_project(project_arn)
delete_test_grid_project(project_arn, params::Dict{String,<:Any})

Deletes a Selenium testing project and all content generated under it. You cannot undo this operation. You cannot delete a project if it has active sessions.

Arguments

project_arn: The ARN of the project to delete, from CreateTestGridProject or ListTestGridProjects.

Main.Device_Farm.delete_upload — Method

delete_upload(arn)
delete_upload(arn, params::Dict{String,<:Any})

Deletes an upload given the upload ARN.

Arguments

arn: Represents the Amazon Resource Name (ARN) of the Device Farm upload to delete.

Main.Device_Farm.delete_vpceconfiguration — Method

delete_vpceconfiguration(arn)
delete_vpceconfiguration(arn, params::Dict{String,<:Any})

Deletes a configuration for your Amazon Virtual Private Cloud (VPC) endpoint.

Arguments

arn: The Amazon Resource Name (ARN) of the VPC endpoint configuration you want to delete.

Main.Device_Farm.get_account_settings — Method

get_account_settings()
get_account_settings(params::Dict{String,<:Any})

Returns the number of unmetered iOS or unmetered Android devices that have been purchased by the account.

Main.Device_Farm.get_device — Method

get_device(arn)
get_device(arn, params::Dict{String,<:Any})

Gets information about a unique device type.

Arguments

arn: The device type's ARN.

Main.Device_Farm.get_device_instance — Method

get_device_instance(arn)
get_device_instance(arn, params::Dict{String,<:Any})

Returns information about a device instance that belongs to a private device fleet.

Arguments

arn: The Amazon Resource Name (ARN) of the instance you're requesting information about.

Main.Device_Farm.get_device_pool — Method

get_device_pool(arn)
get_device_pool(arn, params::Dict{String,<:Any})

Gets information about a device pool.

Arguments

arn: The device pool's ARN.

Main.Device_Farm.get_device_pool_compatibility — Method

get_device_pool_compatibility(device_pool_arn)
get_device_pool_compatibility(device_pool_arn, params::Dict{String,<:Any})

Gets information about compatibility with a device pool.

Arguments

device_pool_arn: The device pool's ARN.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"appArn": The ARN of the app that is associated with the specified device pool.

"configuration": An object that contains information about the settings for a run.

"test": Information about the uploaded test to be run against the device pool.

"testType": The test type for the specified device pool. Allowed values include the following: BUILTINFUZZ. BUILTINEXPLORER. For Android, an app explorer that traverses an Android app, interacting with it and capturing screenshots at the same time. APPIUMJAVAJUNIT. APPIUMJAVATESTNG. APPIUMPYTHON. APPIUMNODE. APPIUMRUBY. APPIUMWEBJAVAJUNIT. APPIUMWEBJAVATESTNG. APPIUMWEBPYTHON. APPIUMWEBNODE. APPIUMWEBRUBY. CALABASH. INSTRUMENTATION. UIAUTOMATION. UIAUTOMATOR. XCTEST. XCTESTUI.

Main.Device_Farm.get_instance_profile — Method

get_instance_profile(arn)
get_instance_profile(arn, params::Dict{String,<:Any})

Returns information about the specified instance profile.

Arguments

arn: The Amazon Resource Name (ARN) of an instance profile.

Main.Device_Farm.get_job — Method

get_job(arn)
get_job(arn, params::Dict{String,<:Any})

Gets information about a job.

Arguments

arn: The job's ARN.

Main.Device_Farm.get_network_profile — Method

get_network_profile(arn)
get_network_profile(arn, params::Dict{String,<:Any})

Returns information about a network profile.

Arguments

arn: The ARN of the network profile to return information about.

Main.Device_Farm.get_offering_status — Method

get_offering_status()
get_offering_status(params::Dict{String,<:Any})

Gets the current status and future status of all offerings purchased by an AWS account. The response indicates how many offerings are currently available and the offerings that will be available in the next period. The API returns a NotEligible error if the user is not permitted to invoke the operation. If you must be able to invoke this operation, contact aws-devicefarm-support@amazon.com.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nextToken": An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Main.Device_Farm.get_project — Method

get_project(arn)
get_project(arn, params::Dict{String,<:Any})

Gets information about a project.

Arguments

arn: The project's ARN.

Main.Device_Farm.get_remote_access_session — Method

get_remote_access_session(arn)
get_remote_access_session(arn, params::Dict{String,<:Any})

Returns a link to a currently running remote access session.

Arguments

arn: The Amazon Resource Name (ARN) of the remote access session about which you want to get session information.

Main.Device_Farm.get_run — Method

get_run(arn)
get_run(arn, params::Dict{String,<:Any})

Gets information about a run.

Arguments

arn: The run's ARN.

Main.Device_Farm.get_suite — Method

get_suite(arn)
get_suite(arn, params::Dict{String,<:Any})

Gets information about a suite.

Arguments

arn: The suite's ARN.

Main.Device_Farm.get_test — Method

get_test(arn)
get_test(arn, params::Dict{String,<:Any})

Gets information about a test.

Arguments

arn: The test's ARN.

Main.Device_Farm.get_test_grid_project — Method

get_test_grid_project(project_arn)
get_test_grid_project(project_arn, params::Dict{String,<:Any})

Retrieves information about a Selenium testing project.

Arguments

project_arn: The ARN of the Selenium testing project, from either CreateTestGridProject or ListTestGridProjects.

Main.Device_Farm.get_test_grid_session — Method

get_test_grid_session()
get_test_grid_session(params::Dict{String,<:Any})

A session is an instance of a browser created through a RemoteWebDriver with the URL from CreateTestGridUrlResulturl. You can use the following to look up sessions: The session ARN (GetTestGridSessionRequestsessionArn). The project ARN and a session ID (GetTestGridSessionRequestprojectArn and GetTestGridSessionRequestsessionId).

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"projectArn": The ARN for the project that this session belongs to. See CreateTestGridProject and ListTestGridProjects.

"sessionArn": An ARN that uniquely identifies a TestGridSession.

"sessionId": An ID associated with this session.

Main.Device_Farm.get_upload — Method

get_upload(arn)
get_upload(arn, params::Dict{String,<:Any})

Gets information about an upload.

Arguments

arn: The upload's ARN.

Main.Device_Farm.get_vpceconfiguration — Method

get_vpceconfiguration(arn)
get_vpceconfiguration(arn, params::Dict{String,<:Any})

Returns information about the configuration settings for your Amazon Virtual Private Cloud (VPC) endpoint.

Arguments

arn: The Amazon Resource Name (ARN) of the VPC endpoint configuration you want to describe.

Main.Device_Farm.install_to_remote_access_session — Method

install_to_remote_access_session(app_arn, remote_access_session_arn)
install_to_remote_access_session(app_arn, remote_access_session_arn, params::Dict{String,<:Any})

Installs an application to the device in a remote access session. For Android applications, the file must be in .apk format. For iOS applications, the file must be in .ipa format.

Arguments

app_arn: The ARN of the app about which you are requesting information.

remote_access_session_arn: The Amazon Resource Name (ARN) of the remote access session about which you are requesting information.

Main.Device_Farm.list_artifacts — Method

list_artifacts(arn, type)
list_artifacts(arn, type, params::Dict{String,<:Any})

Gets information about artifacts.

Arguments

arn: The run, job, suite, or test ARN.

type: The artifacts' type. Allowed values include: FILE LOG SCREENSHOT

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nextToken": An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Main.Device_Farm.list_device_instances — Method

list_device_instances()
list_device_instances(params::Dict{String,<:Any})

Returns information about the private device instances associated with one or more AWS accounts.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": An integer that specifies the maximum number of items you want to return in the API response.

"nextToken": An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Main.Device_Farm.list_device_pools — Method

list_device_pools(arn)
list_device_pools(arn, params::Dict{String,<:Any})

Gets information about device pools.

Arguments

arn: The project ARN.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nextToken": An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

"type": The device pools' type. Allowed values include: CURATED: A device pool that is created and managed by AWS Device Farm. PRIVATE: A device pool that is created and managed by the device pool developer.

Main.Device_Farm.list_devices — Method

list_devices()
list_devices(params::Dict{String,<:Any})

Gets information about unique device types.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"arn": The Amazon Resource Name (ARN) of the project.

"filters": Used to select a set of devices. A filter is made up of an attribute, an operator, and one or more values. Attribute: The aspect of a device such as platform or model used as the selection criteria in a device filter. Allowed values include: ARN: The Amazon Resource Name (ARN) of the device (for example, arn:aws:devicefarm:us-west-2::device:12345Example). PLATFORM: The device platform. Valid values are ANDROID or IOS. OSVERSION: The operating system version (for example, 10.3.2). MODEL: The device model (for example, iPad 5th Gen). AVAILABILITY: The current availability of the device. Valid values are AVAILABLE, HIGHLYAVAILABLE, BUSY, or TEMPORARYNOTAVAILABLE. FORMFACTOR: The device form factor. Valid values are PHONE or TABLET. MANUFACTURER: The device manufacturer (for example, Apple). REMOTEACCESSENABLED: Whether the device is enabled for remote access. Valid values are TRUE or FALSE. REMOTEDEBUGENABLED: Whether the device is enabled for remote debugging. Valid values are TRUE or FALSE. Because remote debugging is no longer supported, this attribute is ignored. INSTANCEARN: The Amazon Resource Name (ARN) of the device instance. INSTANCELABELS: The label of the device instance. FLEETTYPE: The fleet type. Valid values are PUBLIC or PRIVATE. Operator: The filter operator. The EQUALS operator is available for every attribute except INSTANCELABELS. The CONTAINS operator is available for the INSTANCELABELS and MODEL attributes. The IN and NOTIN operators are available for the ARN, OSVERSION, MODEL, MANUFACTURER, and INSTANCEARN attributes. The LESSTHAN, GREATERTHAN, LESSTHANOREQUALS, and GREATERTHANOREQUALS operators are also available for the OSVERSION attribute. Values: An array of one or more filter values. The IN and NOTIN operators take a values array that has one or more elements. The other operators require an array with a single element. In a request, the AVAILABILITY attribute takes the following values: AVAILABLE, HIGHLYAVAILABLE, BUSY, or TEMPORARYNOTAVAILABLE.

"nextToken": An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Main.Device_Farm.list_instance_profiles — Method

list_instance_profiles()
list_instance_profiles(params::Dict{String,<:Any})

Returns information about all the instance profiles in an AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": An integer that specifies the maximum number of items you want to return in the API response.

"nextToken": An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Main.Device_Farm.list_jobs — Method

list_jobs(arn)
list_jobs(arn, params::Dict{String,<:Any})

Gets information about jobs for a given test run.

Arguments

arn: The run's Amazon Resource Name (ARN).

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nextToken": An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Main.Device_Farm.list_network_profiles — Method

list_network_profiles(arn)
list_network_profiles(arn, params::Dict{String,<:Any})

Returns the list of available network profiles.

Arguments

arn: The Amazon Resource Name (ARN) of the project for which you want to list network profiles.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nextToken": An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

"type": The type of network profile to return information about. Valid values are listed here.

Main.Device_Farm.list_offering_promotions — Method

list_offering_promotions()
list_offering_promotions(params::Dict{String,<:Any})

Returns a list of offering promotions. Each offering promotion record contains the ID and description of the promotion. The API returns a NotEligible error if the caller is not permitted to invoke the operation. Contact aws-devicefarm-support@amazon.com if you must be able to invoke this operation.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nextToken": An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Main.Device_Farm.list_offering_transactions — Method

list_offering_transactions()
list_offering_transactions(params::Dict{String,<:Any})

Returns a list of all historical purchases, renewals, and system renewal transactions for an AWS account. The list is paginated and ordered by a descending timestamp (most recent transactions are first). The API returns a NotEligible error if the user is not permitted to invoke the operation. If you must be able to invoke this operation, contact aws-devicefarm-support@amazon.com.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nextToken": An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Main.Device_Farm.list_offerings — Method

list_offerings()
list_offerings(params::Dict{String,<:Any})

Returns a list of products or offerings that the user can manage through the API. Each offering record indicates the recurring price per unit and the frequency for that offering. The API returns a NotEligible error if the user is not permitted to invoke the operation. If you must be able to invoke this operation, contact aws-devicefarm-support@amazon.com.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nextToken": An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Main.Device_Farm.list_projects — Method

list_projects()
list_projects(params::Dict{String,<:Any})

Gets information about projects.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"arn": Optional. If no Amazon Resource Name (ARN) is specified, then AWS Device Farm returns a list of all projects for the AWS account. You can also specify a project ARN.

"nextToken": An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Main.Device_Farm.list_remote_access_sessions — Method

list_remote_access_sessions(arn)
list_remote_access_sessions(arn, params::Dict{String,<:Any})

Returns a list of all currently running remote access sessions.

Arguments

arn: The Amazon Resource Name (ARN) of the project about which you are requesting information.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nextToken": An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Main.Device_Farm.list_runs — Method

list_runs(arn)
list_runs(arn, params::Dict{String,<:Any})

Gets information about runs, given an AWS Device Farm project ARN.

Arguments

arn: The Amazon Resource Name (ARN) of the project for which you want to list runs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nextToken": An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Main.Device_Farm.list_samples — Method

list_samples(arn)
list_samples(arn, params::Dict{String,<:Any})

Gets information about samples, given an AWS Device Farm job ARN.

Arguments

arn: The Amazon Resource Name (ARN) of the job used to list samples.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nextToken": An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Main.Device_Farm.list_suites — Method

list_suites(arn)
list_suites(arn, params::Dict{String,<:Any})

Gets information about test suites for a given job.

Arguments

arn: The job's Amazon Resource Name (ARN).

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nextToken": An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Main.Device_Farm.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

List the tags for an AWS Device Farm resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource or resources for which to list tags. You can associate tags with the following Device Farm resources: PROJECT, RUN, NETWORKPROFILE, INSTANCEPROFILE, DEVICEINSTANCE, SESSION, DEVICEPOOL, DEVICE, and VPCE_CONFIGURATION.

Main.Device_Farm.list_test_grid_projects — Method

list_test_grid_projects()
list_test_grid_projects(params::Dict{String,<:Any})

Gets a list of all Selenium testing projects in your account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResult": Return no more than this number of results.

"nextToken": From a response, used to continue a paginated listing.

Main.Device_Farm.list_test_grid_session_actions — Method

list_test_grid_session_actions(session_arn)
list_test_grid_session_actions(session_arn, params::Dict{String,<:Any})

Returns a list of the actions taken in a TestGridSession.

Arguments

session_arn: The ARN of the session to retrieve.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResult": The maximum number of sessions to return per response.

"nextToken": Pagination token.

Main.Device_Farm.list_test_grid_session_artifacts — Method

list_test_grid_session_artifacts(session_arn)
list_test_grid_session_artifacts(session_arn, params::Dict{String,<:Any})

Retrieves a list of artifacts created during the session.

Arguments

session_arn: The ARN of a TestGridSession.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResult": The maximum number of results to be returned by a request.

"nextToken": Pagination token.

"type": Limit results to a specified type of artifact.

Main.Device_Farm.list_test_grid_sessions — Method

list_test_grid_sessions(project_arn)
list_test_grid_sessions(project_arn, params::Dict{String,<:Any})

Retrieves a list of sessions for a TestGridProject.

Arguments

project_arn: ARN of a TestGridProject.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"creationTimeAfter": Return only sessions created after this time.

"creationTimeBefore": Return only sessions created before this time.

"endTimeAfter": Return only sessions that ended after this time.

"endTimeBefore": Return only sessions that ended before this time.

"maxResult": Return only this many results at a time.

"nextToken": Pagination token.

"status": Return only sessions in this state.

Main.Device_Farm.list_tests — Method

list_tests(arn)
list_tests(arn, params::Dict{String,<:Any})

Gets information about tests in a given test suite.

Arguments

arn: The test suite's Amazon Resource Name (ARN).

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nextToken": An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Main.Device_Farm.list_unique_problems — Method

list_unique_problems(arn)
list_unique_problems(arn, params::Dict{String,<:Any})

Gets information about unique problems, such as exceptions or crashes. Unique problems are defined as a single instance of an error across a run, job, or suite. For example, if a call in your application consistently raises an exception (OutOfBoundsException in MyActivity.java:386), ListUniqueProblems returns a single entry instead of many individual entries for that exception.

Arguments

arn: The unique problems' ARNs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nextToken": An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Main.Device_Farm.list_uploads — Method

list_uploads(arn)
list_uploads(arn, params::Dict{String,<:Any})

Gets information about uploads, given an AWS Device Farm project ARN.

Arguments

arn: The Amazon Resource Name (ARN) of the project for which you want to list uploads.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nextToken": An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

"type": The type of upload. Must be one of the following values: ANDROIDAPP IOSAPP WEBAPP EXTERNALDATA APPIUMJAVAJUNITTESTPACKAGE APPIUMJAVATESTNGTESTPACKAGE APPIUMPYTHONTESTPACKAGE APPIUMNODETESTPACKAGE APPIUMRUBYTESTPACKAGE APPIUMWEBJAVAJUNITTESTPACKAGE APPIUMWEBJAVATESTNGTESTPACKAGE APPIUMWEBPYTHONTESTPACKAGE APPIUMWEBNODETESTPACKAGE APPIUMWEBRUBYTESTPACKAGE CALABASHTESTPACKAGE INSTRUMENTATIONTESTPACKAGE UIAUTOMATIONTESTPACKAGE UIAUTOMATORTESTPACKAGE XCTESTTESTPACKAGE XCTESTUITESTPACKAGE APPIUMJAVAJUNITTESTSPEC APPIUMJAVATESTNGTESTSPEC APPIUMPYTHONTESTSPEC APPIUMNODETESTSPEC APPIUMRUBYTESTSPEC APPIUMWEBJAVAJUNITTESTSPEC APPIUMWEBJAVATESTNGTESTSPEC APPIUMWEBPYTHONTESTSPEC APPIUMWEBNODETESTSPEC APPIUMWEBRUBYTESTSPEC INSTRUMENTATIONTESTSPEC XCTESTUITESTSPEC

Main.Device_Farm.list_vpceconfigurations — Method

list_vpceconfigurations()
list_vpceconfigurations(params::Dict{String,<:Any})

Returns information about all Amazon Virtual Private Cloud (VPC) endpoint configurations in the AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": An integer that specifies the maximum number of items you want to return in the API response.

"nextToken": An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Main.Device_Farm.purchase_offering — Method

purchase_offering(offering_id, quantity)
purchase_offering(offering_id, quantity, params::Dict{String,<:Any})

Immediately purchases offerings for an AWS account. Offerings renew with the latest total purchased quantity for an offering, unless the renewal was overridden. The API returns a NotEligible error if the user is not permitted to invoke the operation. If you must be able to invoke this operation, contact aws-devicefarm-support@amazon.com.

Arguments

offering_id: The ID of the offering.

quantity: The number of device slots to purchase in an offering request.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"offeringPromotionId": The ID of the offering promotion to be applied to the purchase.

Main.Device_Farm.renew_offering — Method

renew_offering(offering_id, quantity)
renew_offering(offering_id, quantity, params::Dict{String,<:Any})

Explicitly sets the quantity of devices to renew for an offering, starting from the effectiveDate of the next period. The API returns a NotEligible error if the user is not permitted to invoke the operation. If you must be able to invoke this operation, contact aws-devicefarm-support@amazon.com.

Arguments

offering_id: The ID of a request to renew an offering.

quantity: The quantity requested in an offering renewal.

Main.Device_Farm.schedule_run — Method

schedule_run(project_arn, test)
schedule_run(project_arn, test, params::Dict{String,<:Any})

Schedules a run.

Arguments

project_arn: The ARN of the project for the run to be scheduled.

test: Information about the test for the run to be scheduled.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"appArn": The ARN of an application package to run tests against, created with CreateUpload. See ListUploads.

"configuration": Information about the settings for the run to be scheduled.

"devicePoolArn": The ARN of the device pool for the run to be scheduled.

"deviceSelectionConfiguration": The filter criteria used to dynamically select a set of devices for a test run and the maximum number of devices to be included in the run. Either devicePoolArn or deviceSelectionConfiguration is required in a request.

"executionConfiguration": Specifies configuration information about a test run, such as the execution timeout (in minutes).

"name": The name for the run to be scheduled.

Main.Device_Farm.stop_job — Method

stop_job(arn)
stop_job(arn, params::Dict{String,<:Any})

Initiates a stop request for the current job. AWS Device Farm immediately stops the job on the device where tests have not started. You are not billed for this device. On the device where tests have started, setup suite and teardown suite tests run to completion on the device. You are billed for setup, teardown, and any tests that were in progress or already completed.

Arguments

arn: Represents the Amazon Resource Name (ARN) of the Device Farm job to stop.

Main.Device_Farm.stop_remote_access_session — Method

stop_remote_access_session(arn)
stop_remote_access_session(arn, params::Dict{String,<:Any})

Ends a specified remote access session.

Arguments

arn: The Amazon Resource Name (ARN) of the remote access session to stop.

Main.Device_Farm.stop_run — Method

stop_run(arn)
stop_run(arn, params::Dict{String,<:Any})

Initiates a stop request for the current test run. AWS Device Farm immediately stops the run on devices where tests have not started. You are not billed for these devices. On devices where tests have started executing, setup suite and teardown suite tests run to completion on those devices. You are billed for setup, teardown, and any tests that were in progress or already completed.

Arguments

arn: Represents the Amazon Resource Name (ARN) of the Device Farm run to stop.

Main.Device_Farm.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Associates the specified tags to a resource with the specified resourceArn. If existing tags on a resource are not specified in the request parameters, they are not changed. When a resource is deleted, the tags associated with that resource are also deleted.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource or resources to which to add tags. You can associate tags with the following Device Farm resources: PROJECT, RUN, NETWORKPROFILE, INSTANCEPROFILE, DEVICEINSTANCE, SESSION, DEVICEPOOL, DEVICE, and VPCE_CONFIGURATION.

tags: The tags to add to the resource. A tag is an array of key-value pairs. Tag keys can have a maximum character length of 128 characters. Tag values can have a maximum length of 256 characters.

Main.Device_Farm.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Deletes the specified tags from a resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource or resources from which to delete tags. You can associate tags with the following Device Farm resources: PROJECT, RUN, NETWORKPROFILE, INSTANCEPROFILE, DEVICEINSTANCE, SESSION, DEVICEPOOL, DEVICE, and VPCE_CONFIGURATION.

tag_keys: The keys of the tags to be removed.

Main.Device_Farm.update_device_instance — Method

update_device_instance(arn)
update_device_instance(arn, params::Dict{String,<:Any})

Updates information about a private device instance.

Arguments

arn: The Amazon Resource Name (ARN) of the device instance.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"labels": An array of strings that you want to associate with the device instance.

"profileArn": The ARN of the profile that you want to associate with the device instance.

Main.Device_Farm.update_device_pool — Method

update_device_pool(arn)
update_device_pool(arn, params::Dict{String,<:Any})

Modifies the name, description, and rules in a device pool given the attributes and the pool ARN. Rule updates are all-or-nothing, meaning they can only be updated as a whole (or not at all).

Arguments

arn: The Amazon Resource Name (ARN) of the Device Farm device pool to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clearMaxDevices": Sets whether the maxDevices parameter applies to your device pool. If you set this parameter to true, the maxDevices parameter does not apply, and Device Farm does not limit the number of devices that it adds to your device pool. In this case, Device Farm adds all available devices that meet the criteria specified in the rules parameter. If you use this parameter in your request, you cannot use the maxDevices parameter in the same request.

"description": A description of the device pool to update.

"maxDevices": The number of devices that Device Farm can add to your device pool. Device Farm adds devices that are available and that meet the criteria that you assign for the rules parameter. Depending on how many devices meet these constraints, your device pool might contain fewer devices than the value for this parameter. By specifying the maximum number of devices, you can control the costs that you incur by running tests. If you use this parameter in your request, you cannot use the clearMaxDevices parameter in the same request.

"name": A string that represents the name of the device pool to update.

"rules": Represents the rules to modify for the device pool. Updating rules is optional. If you update rules for your request, the update replaces the existing rules.

Main.Device_Farm.update_instance_profile — Method

update_instance_profile(arn)
update_instance_profile(arn, params::Dict{String,<:Any})

Updates information about an existing private device instance profile.

Arguments

arn: The Amazon Resource Name (ARN) of the instance profile.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The updated description for your instance profile.

"excludeAppPackagesFromCleanup": An array of strings that specifies the list of app packages that should not be cleaned up from the device after a test run is over. The list of packages is only considered if you set packageCleanup to true.

"name": The updated name for your instance profile.

"packageCleanup": The updated choice for whether you want to specify package cleanup. The default value is false for private devices.

"rebootAfterUse": The updated choice for whether you want to reboot the device after use. The default value is true.

Main.Device_Farm.update_network_profile — Method

update_network_profile(arn)
update_network_profile(arn, params::Dict{String,<:Any})

Updates the network profile.

Arguments

arn: The Amazon Resource Name (ARN) of the project for which you want to update network profile settings.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The description of the network profile about which you are returning information.

"downlinkBandwidthBits": The data throughput rate in bits per second, as an integer from 0 to 104857600.

"downlinkDelayMs": Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

"downlinkJitterMs": Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

"downlinkLossPercent": Proportion of received packets that fail to arrive from 0 to 100 percent.

"name": The name of the network profile about which you are returning information.

"type": The type of network profile to return information about. Valid values are listed here.

"uplinkBandwidthBits": The data throughput rate in bits per second, as an integer from 0 to 104857600.

"uplinkDelayMs": Delay time for all packets to destination in milliseconds as an integer from 0 to 2000.

"uplinkJitterMs": Time variation in the delay of received packets in milliseconds as an integer from 0 to 2000.

"uplinkLossPercent": Proportion of transmitted packets that fail to arrive from 0 to 100 percent.

Main.Device_Farm.update_project — Method

update_project(arn)
update_project(arn, params::Dict{String,<:Any})

Modifies the specified project name, given the project ARN and a new name.

Arguments

arn: The Amazon Resource Name (ARN) of the project whose name to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"defaultJobTimeoutMinutes": The number of minutes a test run in the project executes before it times out.

"name": A string that represents the new name of the project that you are updating.

"vpcConfig": The VPC security groups and subnets that are attached to a project.

Main.Device_Farm.update_test_grid_project — Method

update_test_grid_project(project_arn)
update_test_grid_project(project_arn, params::Dict{String,<:Any})

Change details of a project.

Arguments

project_arn: ARN of the project to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": Human-readable description for the project.

"name": Human-readable name for the project.

"vpcConfig": The VPC security groups and subnets that are attached to a project.

Main.Device_Farm.update_upload — Method

update_upload(arn)
update_upload(arn, params::Dict{String,<:Any})

Updates an uploaded test spec.

Arguments

arn: The Amazon Resource Name (ARN) of the uploaded test spec.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"contentType": The upload's content type (for example, application/x-yaml).

"editContent": Set to true if the YAML file has changed and must be updated. Otherwise, set to false.

"name": The upload's test spec file name. The name must not contain any forward slashes (/). The test spec file name must end with the .yaml or .yml file extension.

Main.Device_Farm.update_vpceconfiguration — Method

update_vpceconfiguration(arn)
update_vpceconfiguration(arn, params::Dict{String,<:Any})

Updates information about an Amazon Virtual Private Cloud (VPC) endpoint configuration.

Arguments

arn: The Amazon Resource Name (ARN) of the VPC endpoint configuration you want to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"serviceDnsName": The DNS (domain) name used to connect to your private service in your VPC. The DNS name must not already be in use on the internet.

"vpceConfigurationDescription": An optional description that provides details about your VPC endpoint configuration.

"vpceConfigurationName": The friendly name you give to your VPC endpoint configuration to manage your configurations more easily.

"vpceServiceName": The name of the VPC endpoint service running in your AWS account that you want Device Farm to test.