Drs

This page documents function available when using the Drs module, created with @service Drs.

Index

Main.Drs.create_extended_source_server

Main.Drs.create_replication_configuration_template

Main.Drs.delete_job

Main.Drs.delete_recovery_instance

Main.Drs.delete_replication_configuration_template

Main.Drs.delete_source_server

Main.Drs.describe_job_log_items

Main.Drs.describe_jobs

Main.Drs.describe_recovery_instances

Main.Drs.describe_recovery_snapshots

Main.Drs.describe_replication_configuration_templates

Main.Drs.describe_source_servers

Main.Drs.disconnect_recovery_instance

Main.Drs.disconnect_source_server

Main.Drs.get_failback_replication_configuration

Main.Drs.get_launch_configuration

Main.Drs.get_replication_configuration

Main.Drs.initialize_service

Main.Drs.list_extensible_source_servers

Main.Drs.list_staging_accounts

Main.Drs.list_tags_for_resource

Main.Drs.retry_data_replication

Main.Drs.reverse_replication

Main.Drs.start_failback_launch

Main.Drs.start_recovery

Main.Drs.start_replication

Main.Drs.stop_failback

Main.Drs.stop_replication

Main.Drs.tag_resource

Main.Drs.terminate_recovery_instances

Main.Drs.untag_resource

Main.Drs.update_failback_replication_configuration

Main.Drs.update_launch_configuration

Main.Drs.update_replication_configuration

Main.Drs.update_replication_configuration_template

Documentation

Main.Drs.create_extended_source_server — Method

create_extended_source_server(source_server_arn)
create_extended_source_server(source_server_arn, params::Dict{String,<:Any})

Create an extended source server in the target Account based on the source server in staging account.

Arguments

source_server_arn: This defines the ARN of the source server in staging Account based on which you want to create an extended source server.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"tags": A list of tags associated with the extended source server.

Main.Drs.create_replication_configuration_template — Method

create_replication_configuration_template(associate_default_security_group, bandwidth_throttling, create_public_ip, data_plane_routing, default_large_staging_disk_type, ebs_encryption, pit_policy, replication_server_instance_type, replication_servers_security_groups_ids, staging_area_subnet_id, staging_area_tags, use_dedicated_replication_server)
create_replication_configuration_template(associate_default_security_group, bandwidth_throttling, create_public_ip, data_plane_routing, default_large_staging_disk_type, ebs_encryption, pit_policy, replication_server_instance_type, replication_servers_security_groups_ids, staging_area_subnet_id, staging_area_tags, use_dedicated_replication_server, params::Dict{String,<:Any})

Creates a new ReplicationConfigurationTemplate.

Arguments

associate_default_security_group: Whether to associate the default Elastic Disaster Recovery Security group with the Replication Configuration Template.

bandwidth_throttling: Configure bandwidth throttling for the outbound data transfer rate of the Source Server in Mbps.

create_public_ip: Whether to create a Public IP for the Recovery Instance by default.

data_plane_routing: The data plane routing mechanism that will be used for replication.

default_large_staging_disk_type: The Staging Disk EBS volume type to be used during replication.

ebs_encryption: The type of EBS encryption to be used during replication.

pit_policy: The Point in time (PIT) policy to manage snapshots taken during replication.

replication_server_instance_type: The instance type to be used for the replication server.

replication_servers_security_groups_ids: The security group IDs that will be used by the replication server.

staging_area_subnet_id: The subnet to be used by the replication staging area.

staging_area_tags: A set of tags to be associated with all resources created in the replication staging area: EC2 replication server, EBS volumes, EBS snapshots, etc.

use_dedicated_replication_server: Whether to use a dedicated Replication Server in the replication staging area.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ebsEncryptionKeyArn": The ARN of the EBS encryption key to be used during replication.

"tags": A set of tags to be associated with the Replication Configuration Template resource.

Main.Drs.delete_job — Method

delete_job(job_id)
delete_job(job_id, params::Dict{String,<:Any})

Deletes a single Job by ID.

Arguments

job_id: The ID of the Job to be deleted.

Main.Drs.delete_recovery_instance — Method

delete_recovery_instance(recovery_instance_id)
delete_recovery_instance(recovery_instance_id, params::Dict{String,<:Any})

Deletes a single Recovery Instance by ID. This deletes the Recovery Instance resource from Elastic Disaster Recovery. The Recovery Instance must be disconnected first in order to delete it.

Arguments

recovery_instance_id: The ID of the Recovery Instance to be deleted.

Main.Drs.delete_replication_configuration_template — Method

delete_replication_configuration_template(replication_configuration_template_id)
delete_replication_configuration_template(replication_configuration_template_id, params::Dict{String,<:Any})

Deletes a single Replication Configuration Template by ID

Arguments

replication_configuration_template_id: The ID of the Replication Configuration Template to be deleted.

Main.Drs.delete_source_server — Method

delete_source_server(source_server_id)
delete_source_server(source_server_id, params::Dict{String,<:Any})

Deletes a single Source Server by ID. The Source Server must be disconnected first.

Arguments

source_server_id: The ID of the Source Server to be deleted.

Main.Drs.describe_job_log_items — Method

describe_job_log_items(job_id)
describe_job_log_items(job_id, params::Dict{String,<:Any})

Retrieves a detailed Job log with pagination.

Arguments

job_id: The ID of the Job for which Job log items will be retrieved.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Maximum number of Job log items to retrieve.

"nextToken": The token of the next Job log items to retrieve.

Main.Drs.describe_jobs — Method

describe_jobs()
describe_jobs(params::Dict{String,<:Any})

Returns a list of Jobs. Use the JobsID and fromDate and toDate filters to limit which jobs are returned. The response is sorted by creationDataTime - latest date first. Jobs are created by the StartRecovery, TerminateRecoveryInstances and StartFailbackLaunch APIs. Jobs are also created by DiagnosticLaunch and TerminateDiagnosticInstances, which are APIs available only to Support and only used in response to relevant support tickets.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filters": A set of filters by which to return Jobs.

"maxResults": Maximum number of Jobs to retrieve.

"nextToken": The token of the next Job to retrieve.

Main.Drs.describe_recovery_instances — Method

describe_recovery_instances()
describe_recovery_instances(params::Dict{String,<:Any})

Lists all Recovery Instances or multiple Recovery Instances by ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filters": A set of filters by which to return Recovery Instances.

"maxResults": Maximum number of Recovery Instances to retrieve.

"nextToken": The token of the next Recovery Instance to retrieve.

Main.Drs.describe_recovery_snapshots — Method

describe_recovery_snapshots(source_server_id)
describe_recovery_snapshots(source_server_id, params::Dict{String,<:Any})

Lists all Recovery Snapshots for a single Source Server.

Arguments

source_server_id: Filter Recovery Snapshots by Source Server ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filters": A set of filters by which to return Recovery Snapshots.

"maxResults": Maximum number of Recovery Snapshots to retrieve.

"nextToken": The token of the next Recovery Snapshot to retrieve.

"order": The sorted ordering by which to return Recovery Snapshots.

Main.Drs.describe_replication_configuration_templates — Method

describe_replication_configuration_templates()
describe_replication_configuration_templates(params::Dict{String,<:Any})

Lists all ReplicationConfigurationTemplates, filtered by Source Server IDs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Maximum number of Replication Configuration Templates to retrieve.

"nextToken": The token of the next Replication Configuration Template to retrieve.

"replicationConfigurationTemplateIDs": The IDs of the Replication Configuration Templates to retrieve. An empty list means all Replication Configuration Templates.

Main.Drs.describe_source_servers — Method

describe_source_servers()
describe_source_servers(params::Dict{String,<:Any})

Lists all Source Servers or multiple Source Servers filtered by ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filters": A set of filters by which to return Source Servers.

"maxResults": Maximum number of Source Servers to retrieve.

"nextToken": The token of the next Source Server to retrieve.

Main.Drs.disconnect_recovery_instance — Method

disconnect_recovery_instance(recovery_instance_id)
disconnect_recovery_instance(recovery_instance_id, params::Dict{String,<:Any})

Disconnect a Recovery Instance from Elastic Disaster Recovery. Data replication is stopped immediately. All AWS resources created by Elastic Disaster Recovery for enabling the replication of the Recovery Instance will be terminated / deleted within 90 minutes. If the agent on the Recovery Instance has not been prevented from communicating with the Elastic Disaster Recovery service, then it will receive a command to uninstall itself (within approximately 10 minutes). The following properties of the Recovery Instance will be changed immediately: dataReplicationInfo.dataReplicationState will be set to DISCONNECTED; The totalStorageBytes property for each of dataReplicationInfo.replicatedDisks will be set to zero; dataReplicationInfo.lagDuration and dataReplicationInfo.lagDuration will be nullified.

Arguments

recovery_instance_id: The ID of the Recovery Instance to disconnect.

Main.Drs.disconnect_source_server — Method

disconnect_source_server(source_server_id)
disconnect_source_server(source_server_id, params::Dict{String,<:Any})

Disconnects a specific Source Server from Elastic Disaster Recovery. Data replication is stopped immediately. All AWS resources created by Elastic Disaster Recovery for enabling the replication of the Source Server will be terminated / deleted within 90 minutes. You cannot disconnect a Source Server if it has a Recovery Instance. If the agent on the Source Server has not been prevented from communicating with the Elastic Disaster Recovery service, then it will receive a command to uninstall itself (within approximately 10 minutes). The following properties of the SourceServer will be changed immediately: dataReplicationInfo.dataReplicationState will be set to DISCONNECTED; The totalStorageBytes property for each of dataReplicationInfo.replicatedDisks will be set to zero; dataReplicationInfo.lagDuration and dataReplicationInfo.lagDuration will be nullified.

Arguments

source_server_id: The ID of the Source Server to disconnect.

Main.Drs.get_failback_replication_configuration — Method

get_failback_replication_configuration(recovery_instance_id)
get_failback_replication_configuration(recovery_instance_id, params::Dict{String,<:Any})

Lists all Failback ReplicationConfigurations, filtered by Recovery Instance ID.

Arguments

recovery_instance_id: The ID of the Recovery Instance whose failback replication configuration should be returned.

Main.Drs.get_launch_configuration — Method

get_launch_configuration(source_server_id)
get_launch_configuration(source_server_id, params::Dict{String,<:Any})

Gets a LaunchConfiguration, filtered by Source Server IDs.

Arguments

source_server_id: The ID of the Source Server that we want to retrieve a Launch Configuration for.

Main.Drs.get_replication_configuration — Method

get_replication_configuration(source_server_id)
get_replication_configuration(source_server_id, params::Dict{String,<:Any})

Gets a ReplicationConfiguration, filtered by Source Server ID.

Arguments

source_server_id: The ID of the Source Serve for this Replication Configuration.r

Main.Drs.initialize_service — Method

initialize_service()
initialize_service(params::Dict{String,<:Any})

Initialize Elastic Disaster Recovery.

Main.Drs.list_extensible_source_servers — Method

list_extensible_source_servers(staging_account_id)
list_extensible_source_servers(staging_account_id, params::Dict{String,<:Any})

Returns a list of source servers on a staging account that are extensible, which means that: a. The source server is not already extended into this Account. b. The source server on the Account we’re reading from is not an extension of another source server.

Arguments

staging_account_id: The Id of the staging Account to retrieve extensible source servers from.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of extensible source servers to retrieve.

"nextToken": The token of the next extensible source server to retrieve.

Main.Drs.list_staging_accounts — Method

list_staging_accounts()
list_staging_accounts(params::Dict{String,<:Any})

Returns an array of staging accounts for existing extended source servers.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of staging Accounts to retrieve.

"nextToken": The token of the next staging Account to retrieve.

Main.Drs.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

List all tags for your Elastic Disaster Recovery resources.

Arguments

resource_arn: The ARN of the resource whose tags should be returned.

Main.Drs.retry_data_replication — Method

retry_data_replication(source_server_id)
retry_data_replication(source_server_id, params::Dict{String,<:Any})

Causes the data replication initiation sequence to begin immediately upon next Handshake for the specified Source Server ID, regardless of when the previous initiation started. This command will work only if the Source Server is stalled or is in a DISCONNECTED or STOPPED state.

Arguments

source_server_id: The ID of the Source Server whose data replication should be retried.

Main.Drs.reverse_replication — Method

reverse_replication(recovery_instance_id)
reverse_replication(recovery_instance_id, params::Dict{String,<:Any})

Start replication to origin / target region - applies only to protected instances that originated in EC2. For recovery instances on target region - starts replication back to origin region. For failback instances on origin region - starts replication to target region to re-protect them.

Arguments

recovery_instance_id: The ID of the Recovery Instance that we want to reverse the replication for.

Main.Drs.start_failback_launch — Method

start_failback_launch(recovery_instance_ids)
start_failback_launch(recovery_instance_ids, params::Dict{String,<:Any})

Initiates a Job for launching the machine that is being failed back to from the specified Recovery Instance. This will run conversion on the failback client and will reboot your machine, thus completing the failback process.

Arguments

recovery_instance_ids: The IDs of the Recovery Instance whose failback launch we want to request.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"tags": The tags to be associated with the failback launch Job.

Main.Drs.start_recovery — Method

start_recovery(source_servers)
start_recovery(source_servers, params::Dict{String,<:Any})

Launches Recovery Instances for the specified Source Servers. For each Source Server you may choose a point in time snapshot to launch from, or use an on demand snapshot.

Arguments

source_servers: The Source Servers that we want to start a Recovery Job for.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"isDrill": Whether this Source Server Recovery operation is a drill or not.

"tags": The tags to be associated with the Recovery Job.

Main.Drs.start_replication — Method

start_replication(source_server_id)
start_replication(source_server_id, params::Dict{String,<:Any})

Starts replication for a stopped Source Server. This action would make the Source Server protected again and restart billing for it.

Arguments

source_server_id: The ID of the Source Server to start replication for.

Main.Drs.stop_failback — Method

stop_failback(recovery_instance_id)
stop_failback(recovery_instance_id, params::Dict{String,<:Any})

Stops the failback process for a specified Recovery Instance. This changes the Failback State of the Recovery Instance back to FAILBACKNOTSTARTED.

Arguments

recovery_instance_id: The ID of the Recovery Instance we want to stop failback for.

Main.Drs.stop_replication — Method

stop_replication(source_server_id)
stop_replication(source_server_id, params::Dict{String,<:Any})

Stops replication for a Source Server. This action would make the Source Server unprotected, delete its existing snapshots and stop billing for it.

Arguments

source_server_id: The ID of the Source Server to stop replication for.

Main.Drs.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Adds or overwrites only the specified tags for the specified Elastic Disaster Recovery resource or resources. When you specify an existing tag key, the value is overwritten with the new value. Each resource can have a maximum of 50 tags. Each tag consists of a key and optional value.

Arguments

resource_arn: ARN of the resource for which tags are to be added or updated.

tags: Array of tags to be added or updated.

Main.Drs.terminate_recovery_instances — Method

terminate_recovery_instances(recovery_instance_ids)
terminate_recovery_instances(recovery_instance_ids, params::Dict{String,<:Any})

Initiates a Job for terminating the EC2 resources associated with the specified Recovery Instances, and then will delete the Recovery Instances from the Elastic Disaster Recovery service.

Arguments

recovery_instance_ids: The IDs of the Recovery Instances that should be terminated.

Main.Drs.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Deletes the specified set of tags from the specified set of Elastic Disaster Recovery resources.

Arguments

resource_arn: ARN of the resource for which tags are to be removed.

tag_keys: Array of tags to be removed.

Main.Drs.update_failback_replication_configuration — Method

update_failback_replication_configuration(recovery_instance_id)
update_failback_replication_configuration(recovery_instance_id, params::Dict{String,<:Any})

Allows you to update the failback replication configuration of a Recovery Instance by ID.

Arguments

recovery_instance_id: The ID of the Recovery Instance.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"bandwidthThrottling": Configure bandwidth throttling for the outbound data transfer rate of the Recovery Instance in Mbps.

"name": The name of the Failback Replication Configuration.

"usePrivateIP": Whether to use Private IP for the failback replication of the Recovery Instance.

Main.Drs.update_launch_configuration — Method

update_launch_configuration(source_server_id)
update_launch_configuration(source_server_id, params::Dict{String,<:Any})

Updates a LaunchConfiguration by Source Server ID.

Arguments

source_server_id: The ID of the Source Server that we want to retrieve a Launch Configuration for.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"copyPrivateIp": Whether we should copy the Private IP of the Source Server to the Recovery Instance.

"copyTags": Whether we want to copy the tags of the Source Server to the EC2 machine of the Recovery Instance.

"launchDisposition": The state of the Recovery Instance in EC2 after the recovery operation.

"licensing": The licensing configuration to be used for this launch configuration.

"name": The name of the launch configuration.

"targetInstanceTypeRightSizingMethod": Whether Elastic Disaster Recovery should try to automatically choose the instance type that best matches the OS, CPU, and RAM of your Source Server.

Main.Drs.update_replication_configuration — Method

update_replication_configuration(source_server_id)
update_replication_configuration(source_server_id, params::Dict{String,<:Any})

Allows you to update a ReplicationConfiguration by Source Server ID.

Arguments

source_server_id: The ID of the Source Server for this Replication Configuration.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"associateDefaultSecurityGroup": Whether to associate the default Elastic Disaster Recovery Security group with the Replication Configuration.

"bandwidthThrottling": Configure bandwidth throttling for the outbound data transfer rate of the Source Server in Mbps.

"createPublicIP": Whether to create a Public IP for the Recovery Instance by default.

"dataPlaneRouting": The data plane routing mechanism that will be used for replication.

"defaultLargeStagingDiskType": The Staging Disk EBS volume type to be used during replication.

"ebsEncryption": The type of EBS encryption to be used during replication.

"ebsEncryptionKeyArn": The ARN of the EBS encryption key to be used during replication.

"name": The name of the Replication Configuration.

"pitPolicy": The Point in time (PIT) policy to manage snapshots taken during replication.

"replicatedDisks": The configuration of the disks of the Source Server to be replicated.

"replicationServerInstanceType": The instance type to be used for the replication server.

"replicationServersSecurityGroupsIDs": The security group IDs that will be used by the replication server.

"stagingAreaSubnetId": The subnet to be used by the replication staging area.

"stagingAreaTags": A set of tags to be associated with all resources created in the replication staging area: EC2 replication server, EBS volumes, EBS snapshots, etc.

"useDedicatedReplicationServer": Whether to use a dedicated Replication Server in the replication staging area.

Main.Drs.update_replication_configuration_template — Method

update_replication_configuration_template(replication_configuration_template_id)
update_replication_configuration_template(replication_configuration_template_id, params::Dict{String,<:Any})

Updates a ReplicationConfigurationTemplate by ID.

Arguments

replication_configuration_template_id: The Replication Configuration Template ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"arn": The Replication Configuration Template ARN.

"associateDefaultSecurityGroup": Whether to associate the default Elastic Disaster Recovery Security group with the Replication Configuration Template.

"bandwidthThrottling": Configure bandwidth throttling for the outbound data transfer rate of the Source Server in Mbps.

"createPublicIP": Whether to create a Public IP for the Recovery Instance by default.

"dataPlaneRouting": The data plane routing mechanism that will be used for replication.

"defaultLargeStagingDiskType": The Staging Disk EBS volume type to be used during replication.

"ebsEncryption": The type of EBS encryption to be used during replication.

"ebsEncryptionKeyArn": The ARN of the EBS encryption key to be used during replication.

"pitPolicy": The Point in time (PIT) policy to manage snapshots taken during replication.

"replicationServerInstanceType": The instance type to be used for the replication server.

"replicationServersSecurityGroupsIDs": The security group IDs that will be used by the replication server.

"stagingAreaSubnetId": The subnet to be used by the replication staging area.

"stagingAreaTags": A set of tags to be associated with all resources created in the replication staging area: EC2 replication server, EBS volumes, EBS snapshots, etc.

"useDedicatedReplicationServer": Whether to use a dedicated Replication Server in the replication staging area.