Ecr

This page documents function available when using the Ecr module, created with @service Ecr.

Index

Main.Ecr.batch_check_layer_availability

Main.Ecr.batch_delete_image

Main.Ecr.batch_get_image

Main.Ecr.batch_get_repository_scanning_configuration

Main.Ecr.complete_layer_upload

Main.Ecr.create_pull_through_cache_rule

Main.Ecr.create_repository

Main.Ecr.create_repository_creation_template

Main.Ecr.delete_lifecycle_policy

Main.Ecr.delete_pull_through_cache_rule

Main.Ecr.delete_registry_policy

Main.Ecr.delete_repository

Main.Ecr.delete_repository_creation_template

Main.Ecr.delete_repository_policy

Main.Ecr.describe_image_replication_status

Main.Ecr.describe_image_scan_findings

Main.Ecr.describe_images

Main.Ecr.describe_pull_through_cache_rules

Main.Ecr.describe_registry

Main.Ecr.describe_repositories

Main.Ecr.describe_repository_creation_templates

Main.Ecr.get_account_setting

Main.Ecr.get_authorization_token

Main.Ecr.get_download_url_for_layer

Main.Ecr.get_lifecycle_policy

Main.Ecr.get_lifecycle_policy_preview

Main.Ecr.get_registry_policy

Main.Ecr.get_registry_scanning_configuration

Main.Ecr.get_repository_policy

Main.Ecr.initiate_layer_upload

Main.Ecr.list_images

Main.Ecr.list_tags_for_resource

Main.Ecr.put_account_setting

Main.Ecr.put_image

Main.Ecr.put_image_scanning_configuration

Main.Ecr.put_image_tag_mutability

Main.Ecr.put_lifecycle_policy

Main.Ecr.put_registry_policy

Main.Ecr.put_registry_scanning_configuration

Main.Ecr.put_replication_configuration

Main.Ecr.set_repository_policy

Main.Ecr.start_image_scan

Main.Ecr.start_lifecycle_policy_preview

Main.Ecr.tag_resource

Main.Ecr.untag_resource

Main.Ecr.update_pull_through_cache_rule

Main.Ecr.update_repository_creation_template

Main.Ecr.upload_layer_part

Main.Ecr.validate_pull_through_cache_rule

Documentation

Main.Ecr.batch_check_layer_availability — Function

batch_check_layer_availability(layer_digests, repository_name)
batch_check_layer_availability(layer_digests, repository_name, params::Dict{String,<:Any})

Checks the availability of one or more image layers in a repository. When an image is pushed to a repository, each image layer is checked to verify if it has been uploaded before. If it has been uploaded, then the image layer is skipped. This operation is used by the Amazon ECR proxy and is not generally used by customers for pulling and pushing images. In most cases, you should use the docker CLI to pull, tag, and push images.

Arguments

layer_digests: The digests of the image layers to check.

repository_name: The name of the repository that is associated with the image layers to check.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"registryId": The Amazon Web Services account ID associated with the registry that contains the image layers to check. If you do not specify a registry, the default registry is assumed.

Main.Ecr.batch_delete_image — Function

batch_delete_image(image_ids, repository_name)
batch_delete_image(image_ids, repository_name, params::Dict{String,<:Any})

Deletes a list of specified images within a repository. Images are specified with either an imageTag or imageDigest. You can remove a tag from an image by specifying the image's tag in your request. When you remove the last tag from an image, the image is deleted from your repository. You can completely delete an image (and all of its tags) by specifying the image's digest in your request.

Arguments

image_ids: A list of image ID references that correspond to images to delete. The format of the imageIds reference is imageTag=tag or imageDigest=digest.

repository_name: The repository that contains the image to delete.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"registryId": The Amazon Web Services account ID associated with the registry that contains the image to delete. If you do not specify a registry, the default registry is assumed.

Main.Ecr.batch_get_image — Function

batch_get_image(image_ids, repository_name)
batch_get_image(image_ids, repository_name, params::Dict{String,<:Any})

Gets detailed information for an image. Images are specified with either an imageTag or imageDigest. When an image is pulled, the BatchGetImage API is called once to retrieve the image manifest.

Arguments

image_ids: A list of image ID references that correspond to images to describe. The format of the imageIds reference is imageTag=tag or imageDigest=digest.

repository_name: The repository that contains the images to describe.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"acceptedMediaTypes": The accepted media types for the request. Valid values: application/vnd.docker.distribution.manifest.v1+json | application/vnd.docker.distribution.manifest.v2+json | application/vnd.oci.image.manifest.v1+json

"registryId": The Amazon Web Services account ID associated with the registry that contains the images to describe. If you do not specify a registry, the default registry is assumed.

Main.Ecr.batch_get_repository_scanning_configuration — Function

batch_get_repository_scanning_configuration(repository_names)
batch_get_repository_scanning_configuration(repository_names, params::Dict{String,<:Any})

Gets the scanning configuration for one or more repositories.

Arguments

repository_names: One or more repository names to get the scanning configuration for.

Main.Ecr.complete_layer_upload — Function

complete_layer_upload(layer_digests, repository_name, upload_id)
complete_layer_upload(layer_digests, repository_name, upload_id, params::Dict{String,<:Any})

Informs Amazon ECR that the image layer upload has completed for a specified registry, repository name, and upload ID. You can optionally provide a sha256 digest of the image layer for data validation purposes. When an image is pushed, the CompleteLayerUpload API is called once per each new image layer to verify that the upload has completed. This operation is used by the Amazon ECR proxy and is not generally used by customers for pulling and pushing images. In most cases, you should use the docker CLI to pull, tag, and push images.

Arguments

layer_digests: The sha256 digest of the image layer.

repository_name: The name of the repository to associate with the image layer.

upload_id: The upload ID from a previous InitiateLayerUpload operation to associate with the image layer.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"registryId": The Amazon Web Services account ID associated with the registry to which to upload layers. If you do not specify a registry, the default registry is assumed.

Main.Ecr.create_pull_through_cache_rule — Function

create_pull_through_cache_rule(ecr_repository_prefix, upstream_registry_url)
create_pull_through_cache_rule(ecr_repository_prefix, upstream_registry_url, params::Dict{String,<:Any})

Creates a pull through cache rule. A pull through cache rule provides a way to cache images from an upstream registry source in your Amazon ECR private registry. For more information, see Using pull through cache rules in the Amazon Elastic Container Registry User Guide.

Arguments

ecr_repository_prefix: The repository name prefix to use when caching images from the source registry.

upstream_registry_url: The registry URL of the upstream public registry to use as the source for the pull through cache rule. The following is the syntax to use for each supported upstream registry. Amazon ECR Public (ecr-public) - public.ecr.aws Docker Hub (docker-hub) - registry-1.docker.io Quay (quay) - quay.io Kubernetes (k8s) - registry.k8s.io GitHub Container Registry (github-container-registry) - ghcr.io Microsoft Azure Container Registry (azure-container-registry) - &lt;custom&gt;.azurecr.io

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"credentialArn": The Amazon Resource Name (ARN) of the Amazon Web Services Secrets Manager secret that identifies the credentials to authenticate to the upstream registry.

"registryId": The Amazon Web Services account ID associated with the registry to create the pull through cache rule for. If you do not specify a registry, the default registry is assumed.

"upstreamRegistry": The name of the upstream registry.

Main.Ecr.create_repository — Function

create_repository(repository_name)
create_repository(repository_name, params::Dict{String,<:Any})

Creates a repository. For more information, see Amazon ECR repositories in the Amazon Elastic Container Registry User Guide.

Arguments

repository_name: The name to use for the repository. The repository name may be specified on its own (such as nginx-web-app) or it can be prepended with a namespace to group the repository into a category (such as project-a/nginx-web-app). The repository name must start with a letter and can only contain lowercase letters, numbers, hyphens, underscores, and forward slashes.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"encryptionConfiguration": The encryption configuration for the repository. This determines how the contents of your repository are encrypted at rest.

"imageScanningConfiguration": The image scanning configuration for the repository. This determines whether images are scanned for known vulnerabilities after being pushed to the repository.

"imageTagMutability": The tag mutability setting for the repository. If this parameter is omitted, the default setting of MUTABLE will be used which will allow image tags to be overwritten. If IMMUTABLE is specified, all image tags within the repository will be immutable which will prevent them from being overwritten.

"registryId": The Amazon Web Services account ID associated with the registry to create the repository. If you do not specify a registry, the default registry is assumed.

"tags": The metadata that you apply to the repository to help you categorize and organize them. Each tag consists of a key and an optional value, both of which you define. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.

Main.Ecr.create_repository_creation_template — Function

create_repository_creation_template(applied_for, prefix)
create_repository_creation_template(applied_for, prefix, params::Dict{String,<:Any})

Creates a repository creation template. This template is used to define the settings for repositories created by Amazon ECR on your behalf. For example, repositories created through pull through cache actions. For more information, see Private repository creation templates in the Amazon Elastic Container Registry User Guide.

Arguments

applied_for: A list of enumerable strings representing the Amazon ECR repository creation scenarios that this template will apply towards. The two supported scenarios are PULLTHROUGHCACHE and REPLICATION

prefix: The repository namespace prefix to associate with the template. All repositories created using this namespace prefix will have the settings defined in this template applied. For example, a prefix of prod would apply to all repositories beginning with prod/. Similarly, a prefix of prod/team would apply to all repositories beginning with prod/team/. To apply a template to all repositories in your registry that don't have an associated creation template, you can use ROOT as the prefix. There is always an assumed / applied to the end of the prefix. If you specify ecr-public as the prefix, Amazon ECR treats that as ecr-public/. When using a pull through cache rule, the repository prefix you specify during rule creation is what you should specify as your repository creation template prefix as well.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"customRoleArn": The ARN of the role to be assumed by Amazon ECR. This role must be in the same account as the registry that you are configuring. Amazon ECR will assume your supplied role when the customRoleArn is specified. When this field isn't specified, Amazon ECR will use the service-linked role for the repository creation template.

"description": A description for the repository creation template.

"encryptionConfiguration": The encryption configuration to use for repositories created using the template.

"imageTagMutability": The tag mutability setting for the repository. If this parameter is omitted, the default setting of MUTABLE will be used which will allow image tags to be overwritten. If IMMUTABLE is specified, all image tags within the repository will be immutable which will prevent them from being overwritten.

"lifecyclePolicy": The lifecycle policy to use for repositories created using the template.

"repositoryPolicy": The repository policy to apply to repositories created using the template. A repository policy is a permissions policy associated with a repository to control access permissions.

"resourceTags": The metadata to apply to the repository to help you categorize and organize. Each tag consists of a key and an optional value, both of which you define. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.

Main.Ecr.delete_lifecycle_policy — Function

delete_lifecycle_policy(repository_name)
delete_lifecycle_policy(repository_name, params::Dict{String,<:Any})

Deletes the lifecycle policy associated with the specified repository.

Arguments

repository_name: The name of the repository.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"registryId": The Amazon Web Services account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

Main.Ecr.delete_pull_through_cache_rule — Function

delete_pull_through_cache_rule(ecr_repository_prefix)
delete_pull_through_cache_rule(ecr_repository_prefix, params::Dict{String,<:Any})

Deletes a pull through cache rule.

Arguments

ecr_repository_prefix: The Amazon ECR repository prefix associated with the pull through cache rule to delete.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"registryId": The Amazon Web Services account ID associated with the registry that contains the pull through cache rule. If you do not specify a registry, the default registry is assumed.

Main.Ecr.delete_registry_policy — Function

delete_registry_policy()
delete_registry_policy(params::Dict{String,<:Any})

Deletes the registry permissions policy.

Main.Ecr.delete_repository — Function

delete_repository(repository_name)
delete_repository(repository_name, params::Dict{String,<:Any})

Deletes a repository. If the repository isn't empty, you must either delete the contents of the repository or use the force option to delete the repository and have Amazon ECR delete all of its contents on your behalf.

Arguments

repository_name: The name of the repository to delete.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"force": If true, deleting the repository force deletes the contents of the repository. If false, the repository must be empty before attempting to delete it.

"registryId": The Amazon Web Services account ID associated with the registry that contains the repository to delete. If you do not specify a registry, the default registry is assumed.

Main.Ecr.delete_repository_creation_template — Function

delete_repository_creation_template(prefix)
delete_repository_creation_template(prefix, params::Dict{String,<:Any})

Deletes a repository creation template.

Arguments

prefix: The repository namespace prefix associated with the repository creation template.

Main.Ecr.delete_repository_policy — Function

delete_repository_policy(repository_name)
delete_repository_policy(repository_name, params::Dict{String,<:Any})

Deletes the repository policy associated with the specified repository.

Arguments

repository_name: The name of the repository that is associated with the repository policy to delete.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"registryId": The Amazon Web Services account ID associated with the registry that contains the repository policy to delete. If you do not specify a registry, the default registry is assumed.

Main.Ecr.describe_image_replication_status — Function

describe_image_replication_status(image_id, repository_name)
describe_image_replication_status(image_id, repository_name, params::Dict{String,<:Any})

Returns the replication status for a specified image.

Arguments

image_id:

repository_name: The name of the repository that the image is in.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"registryId": The Amazon Web Services account ID associated with the registry. If you do not specify a registry, the default registry is assumed.

Main.Ecr.describe_image_scan_findings — Function

describe_image_scan_findings(image_id, repository_name)
describe_image_scan_findings(image_id, repository_name, params::Dict{String,<:Any})

Returns the scan findings for the specified image.

Arguments

image_id:

repository_name: The repository for the image for which to describe the scan findings.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of image scan results returned by DescribeImageScanFindings in paginated output. When this parameter is used, DescribeImageScanFindings only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeImageScanFindings request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then DescribeImageScanFindings returns up to 100 results and a nextToken value, if applicable.

"nextToken": The nextToken value returned from a previous paginated DescribeImageScanFindings request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

"registryId": The Amazon Web Services account ID associated with the registry that contains the repository in which to describe the image scan findings for. If you do not specify a registry, the default registry is assumed.

Main.Ecr.describe_images — Function

describe_images(repository_name)
describe_images(repository_name, params::Dict{String,<:Any})

Returns metadata about the images in a repository. Beginning with Docker version 1.9, the Docker client compresses image layers before pushing them to a V2 Docker registry. The output of the docker images command shows the uncompressed image size, so it may return a larger image size than the image sizes returned by DescribeImages.

Arguments

repository_name: The repository that contains the images to describe.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filter": The filter key and value with which to filter your DescribeImages results.

"imageIds": The list of image IDs for the requested repository.

"maxResults": The maximum number of repository results returned by DescribeImages in paginated output. When this parameter is used, DescribeImages only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeImages request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then DescribeImages returns up to 100 results and a nextToken value, if applicable. This option cannot be used when you specify images with imageIds.

"nextToken": The nextToken value returned from a previous paginated DescribeImages request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return. This option cannot be used when you specify images with imageIds.

"registryId": The Amazon Web Services account ID associated with the registry that contains the repository in which to describe images. If you do not specify a registry, the default registry is assumed.

Main.Ecr.describe_pull_through_cache_rules — Function

describe_pull_through_cache_rules()
describe_pull_through_cache_rules(params::Dict{String,<:Any})

Returns the pull through cache rules for a registry.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ecrRepositoryPrefixes": The Amazon ECR repository prefixes associated with the pull through cache rules to return. If no repository prefix value is specified, all pull through cache rules are returned.

"maxResults": The maximum number of pull through cache rules returned by DescribePullThroughCacheRulesRequest in paginated output. When this parameter is used, DescribePullThroughCacheRulesRequest only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribePullThroughCacheRulesRequest request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then DescribePullThroughCacheRulesRequest returns up to 100 results and a nextToken value, if applicable.

"nextToken": The nextToken value returned from a previous paginated DescribePullThroughCacheRulesRequest request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return.

"registryId": The Amazon Web Services account ID associated with the registry to return the pull through cache rules for. If you do not specify a registry, the default registry is assumed.

Main.Ecr.describe_registry — Function

describe_registry()
describe_registry(params::Dict{String,<:Any})

Describes the settings for a registry. The replication configuration for a repository can be created or updated with the PutReplicationConfiguration API action.

Main.Ecr.describe_repositories — Function

describe_repositories()
describe_repositories(params::Dict{String,<:Any})

Describes image repositories in a registry.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of repository results returned by DescribeRepositories in paginated output. When this parameter is used, DescribeRepositories only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeRepositories request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then DescribeRepositories returns up to 100 results and a nextToken value, if applicable. This option cannot be used when you specify repositories with repositoryNames.

"nextToken": The nextToken value returned from a previous paginated DescribeRepositories request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return. This option cannot be used when you specify repositories with repositoryNames. This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

"registryId": The Amazon Web Services account ID associated with the registry that contains the repositories to be described. If you do not specify a registry, the default registry is assumed.

"repositoryNames": A list of repositories to describe. If this parameter is omitted, then all repositories in a registry are described.

Main.Ecr.describe_repository_creation_templates — Function

describe_repository_creation_templates()
describe_repository_creation_templates(params::Dict{String,<:Any})

Returns details about the repository creation templates in a registry. The prefixes request parameter can be used to return the details for a specific repository creation template.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of repository results returned by DescribeRepositoryCreationTemplatesRequest in paginated output. When this parameter is used, DescribeRepositoryCreationTemplatesRequest only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another DescribeRepositoryCreationTemplatesRequest request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then DescribeRepositoryCreationTemplatesRequest returns up to 100 results and a nextToken value, if applicable.

"nextToken": The nextToken value returned from a previous paginated DescribeRepositoryCreationTemplates request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return. This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

"prefixes": The repository namespace prefixes associated with the repository creation templates to describe. If this value is not specified, all repository creation templates are returned.

Main.Ecr.get_account_setting — Function

get_account_setting(name)
get_account_setting(name, params::Dict{String,<:Any})

Retrieves the basic scan type version name.

Arguments

name: Basic scan type version name.

Main.Ecr.get_authorization_token — Function

get_authorization_token()
get_authorization_token(params::Dict{String,<:Any})

Retrieves an authorization token. An authorization token represents your IAM authentication credentials and can be used to access any Amazon ECR registry that your IAM principal has access to. The authorization token is valid for 12 hours. The authorizationToken returned is a base64 encoded string that can be decoded and used in a docker login command to authenticate to a registry. The CLI offers an get-login-password command that simplifies the login process. For more information, see Registry authentication in the Amazon Elastic Container Registry User Guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"registryIds": A list of Amazon Web Services account IDs that are associated with the registries for which to get AuthorizationData objects. If you do not specify a registry, the default registry is assumed.

Main.Ecr.get_download_url_for_layer — Function

get_download_url_for_layer(layer_digest, repository_name)
get_download_url_for_layer(layer_digest, repository_name, params::Dict{String,<:Any})

Retrieves the pre-signed Amazon S3 download URL corresponding to an image layer. You can only get URLs for image layers that are referenced in an image. When an image is pulled, the GetDownloadUrlForLayer API is called once per image layer that is not already cached. This operation is used by the Amazon ECR proxy and is not generally used by customers for pulling and pushing images. In most cases, you should use the docker CLI to pull, tag, and push images.

Arguments

layer_digest: The digest of the image layer to download.

repository_name: The name of the repository that is associated with the image layer to download.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"registryId": The Amazon Web Services account ID associated with the registry that contains the image layer to download. If you do not specify a registry, the default registry is assumed.

Main.Ecr.get_lifecycle_policy — Function

get_lifecycle_policy(repository_name)
get_lifecycle_policy(repository_name, params::Dict{String,<:Any})

Retrieves the lifecycle policy for the specified repository.

Arguments

repository_name: The name of the repository.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"registryId": The Amazon Web Services account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

Main.Ecr.get_lifecycle_policy_preview — Function

get_lifecycle_policy_preview(repository_name)
get_lifecycle_policy_preview(repository_name, params::Dict{String,<:Any})

Retrieves the results of the lifecycle policy preview request for the specified repository.

Arguments

repository_name: The name of the repository.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filter": An optional parameter that filters results based on image tag status and all tags, if tagged.

"imageIds": The list of imageIDs to be included.

"maxResults": The maximum number of repository results returned by GetLifecyclePolicyPreviewRequest in&#x2028; paginated output. When this parameter is used, GetLifecyclePolicyPreviewRequest only returns&#x2028; maxResults results in a single page along with a nextToken&#x2028; response element. The remaining results of the initial request can be seen by sending&#x2028; another GetLifecyclePolicyPreviewRequest request with the returned nextToken&#x2028; value. This value can be between 1 and 1000. If this&#x2028; parameter is not used, then GetLifecyclePolicyPreviewRequest returns up to&#x2028; 100 results and a nextToken value, if&#x2028; applicable. This option cannot be used when you specify images with imageIds.

"nextToken": The nextToken value returned from a previous paginated&#x2028; GetLifecyclePolicyPreviewRequest request where maxResults was used and the&#x2028; results exceeded the value of that parameter. Pagination continues from the end of the&#x2028; previous results that returned the nextToken value. This value is&#x2028; null when there are no more results to return. This option cannot be used when you specify images with imageIds.

"registryId": The Amazon Web Services account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

Main.Ecr.get_registry_policy — Function

get_registry_policy()
get_registry_policy(params::Dict{String,<:Any})

Retrieves the permissions policy for a registry.

Main.Ecr.get_registry_scanning_configuration — Function

get_registry_scanning_configuration()
get_registry_scanning_configuration(params::Dict{String,<:Any})

Retrieves the scanning configuration for a registry.

Main.Ecr.get_repository_policy — Function

get_repository_policy(repository_name)
get_repository_policy(repository_name, params::Dict{String,<:Any})

Retrieves the repository policy for the specified repository.

Arguments

repository_name: The name of the repository with the policy to retrieve.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"registryId": The Amazon Web Services account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

Main.Ecr.initiate_layer_upload — Function

initiate_layer_upload(repository_name)
initiate_layer_upload(repository_name, params::Dict{String,<:Any})

Notifies Amazon ECR that you intend to upload an image layer. When an image is pushed, the InitiateLayerUpload API is called once per image layer that has not already been uploaded. Whether or not an image layer has been uploaded is determined by the BatchCheckLayerAvailability API action. This operation is used by the Amazon ECR proxy and is not generally used by customers for pulling and pushing images. In most cases, you should use the docker CLI to pull, tag, and push images.

Arguments

repository_name: The name of the repository to which you intend to upload layers.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"registryId": The Amazon Web Services account ID associated with the registry to which you intend to upload layers. If you do not specify a registry, the default registry is assumed.

Main.Ecr.list_images — Function

list_images(repository_name)
list_images(repository_name, params::Dict{String,<:Any})

Lists all the image IDs for the specified repository. You can filter images based on whether or not they are tagged by using the tagStatus filter and specifying either TAGGED, UNTAGGED or ANY. For example, you can filter your results to return only UNTAGGED images and then pipe that result to a BatchDeleteImage operation to delete them. Or, you can filter your results to return only TAGGED images to list all of the tags in your repository.

Arguments

repository_name: The repository with image IDs to be listed.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filter": The filter key and value with which to filter your ListImages results.

"maxResults": The maximum number of image results returned by ListImages in paginated output. When this parameter is used, ListImages only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListImages request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then ListImages returns up to 100 results and a nextToken value, if applicable.

"nextToken": The nextToken value returned from a previous paginated ListImages request where maxResults was used and the results exceeded the value of that parameter. Pagination continues from the end of the previous results that returned the nextToken value. This value is null when there are no more results to return. This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

"registryId": The Amazon Web Services account ID associated with the registry that contains the repository in which to list images. If you do not specify a registry, the default registry is assumed.

Main.Ecr.list_tags_for_resource — Function

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

List the tags for an Amazon ECR resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) that identifies the resource for which to list the tags. Currently, the only supported resource is an Amazon ECR repository.

Main.Ecr.put_account_setting — Function

put_account_setting(name, value)
put_account_setting(name, value, params::Dict{String,<:Any})

Allows you to change the basic scan type version by setting the name parameter to either CLAIR to AWS_NATIVE.

Arguments

name: Basic scan type version name.

value: Setting value that determines what basic scan type is being used: AWS_NATIVE or CLAIR.

Main.Ecr.put_image — Function

put_image(image_manifest, repository_name)
put_image(image_manifest, repository_name, params::Dict{String,<:Any})

Creates or updates the image manifest and tags associated with an image. When an image is pushed and all new image layers have been uploaded, the PutImage API is called once to create or update the image manifest and the tags associated with the image. This operation is used by the Amazon ECR proxy and is not generally used by customers for pulling and pushing images. In most cases, you should use the docker CLI to pull, tag, and push images.

Arguments

image_manifest: The image manifest corresponding to the image to be uploaded.

repository_name: The name of the repository in which to put the image.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"imageDigest": The image digest of the image manifest corresponding to the image.

"imageManifestMediaType": The media type of the image manifest. If you push an image manifest that does not contain the mediaType field, you must specify the imageManifestMediaType in the request.

"imageTag": The tag to associate with the image. This parameter is required for images that use the Docker Image Manifest V2 Schema 2 or Open Container Initiative (OCI) formats.

"registryId": The Amazon Web Services account ID associated with the registry that contains the repository in which to put the image. If you do not specify a registry, the default registry is assumed.

Main.Ecr.put_image_scanning_configuration — Function

put_image_scanning_configuration(image_scanning_configuration, repository_name)
put_image_scanning_configuration(image_scanning_configuration, repository_name, params::Dict{String,<:Any})

The PutImageScanningConfiguration API is being deprecated, in favor of specifying the image scanning configuration at the registry level. For more information, see PutRegistryScanningConfiguration. Updates the image scanning configuration for the specified repository.

Arguments

image_scanning_configuration: The image scanning configuration for the repository. This setting determines whether images are scanned for known vulnerabilities after being pushed to the repository.

repository_name: The name of the repository in which to update the image scanning configuration setting.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"registryId": The Amazon Web Services account ID associated with the registry that contains the repository in which to update the image scanning configuration setting. If you do not specify a registry, the default registry is assumed.

Main.Ecr.put_image_tag_mutability — Function

put_image_tag_mutability(image_tag_mutability, repository_name)
put_image_tag_mutability(image_tag_mutability, repository_name, params::Dict{String,<:Any})

Updates the image tag mutability settings for the specified repository. For more information, see Image tag mutability in the Amazon Elastic Container Registry User Guide.

Arguments

image_tag_mutability: The tag mutability setting for the repository. If MUTABLE is specified, image tags can be overwritten. If IMMUTABLE is specified, all image tags within the repository will be immutable which will prevent them from being overwritten.

repository_name: The name of the repository in which to update the image tag mutability settings.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"registryId": The Amazon Web Services account ID associated with the registry that contains the repository in which to update the image tag mutability settings. If you do not specify a registry, the default registry is assumed.

Main.Ecr.put_lifecycle_policy — Function

put_lifecycle_policy(lifecycle_policy_text, repository_name)
put_lifecycle_policy(lifecycle_policy_text, repository_name, params::Dict{String,<:Any})

Creates or updates the lifecycle policy for the specified repository. For more information, see Lifecycle policy template.

Arguments

lifecycle_policy_text: The JSON repository policy text to apply to the repository.

repository_name: The name of the repository to receive the policy.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"registryId": The Amazon Web Services account ID associated with the registry that contains the repository. If you do&#x2028; not specify a registry, the default registry is assumed.

Main.Ecr.put_registry_policy — Function

put_registry_policy(policy_text)
put_registry_policy(policy_text, params::Dict{String,<:Any})

Creates or updates the permissions policy for your registry. A registry policy is used to specify permissions for another Amazon Web Services account and is used when configuring cross-account replication. For more information, see Registry permissions in the Amazon Elastic Container Registry User Guide.

Arguments

policy_text: The JSON policy text to apply to your registry. The policy text follows the same format as IAM policy text. For more information, see Registry permissions in the Amazon Elastic Container Registry User Guide.

Main.Ecr.put_registry_scanning_configuration — Function

put_registry_scanning_configuration()
put_registry_scanning_configuration(params::Dict{String,<:Any})

Creates or updates the scanning configuration for your private registry.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"rules": The scanning rules to use for the registry. A scanning rule is used to determine which repository filters are used and at what frequency scanning will occur.

"scanType": The scanning type to set for the registry. When a registry scanning configuration is not defined, by default the BASIC scan type is used. When basic scanning is used, you may specify filters to determine which individual repositories, or all repositories, are scanned when new images are pushed to those repositories. Alternatively, you can do manual scans of images with basic scanning. When the ENHANCED scan type is set, Amazon Inspector provides automated vulnerability scanning. You may choose between continuous scanning or scan on push and you may specify filters to determine which individual repositories, or all repositories, are scanned.

Main.Ecr.put_replication_configuration — Function

put_replication_configuration(replication_configuration)
put_replication_configuration(replication_configuration, params::Dict{String,<:Any})

Creates or updates the replication configuration for a registry. The existing replication configuration for a repository can be retrieved with the DescribeRegistry API action. The first time the PutReplicationConfiguration API is called, a service-linked IAM role is created in your account for the replication process. For more information, see Using service-linked roles for Amazon ECR in the Amazon Elastic Container Registry User Guide. For more information on the custom role for replication, see Creating an IAM role for replication. When configuring cross-account replication, the destination account must grant the source account permission to replicate. This permission is controlled using a registry permissions policy. For more information, see PutRegistryPolicy.

Arguments

replication_configuration: An object representing the replication configuration for a registry.

Main.Ecr.set_repository_policy — Function

set_repository_policy(policy_text, repository_name)
set_repository_policy(policy_text, repository_name, params::Dict{String,<:Any})

Applies a repository policy to the specified repository to control access permissions. For more information, see Amazon ECR Repository policies in the Amazon Elastic Container Registry User Guide.

Arguments

policy_text: The JSON repository policy text to apply to the repository. For more information, see Amazon ECR repository policies in the Amazon Elastic Container Registry User Guide.

repository_name: The name of the repository to receive the policy.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"force": If the policy you are attempting to set on a repository policy would prevent you from setting another policy in the future, you must force the SetRepositoryPolicy operation. This is intended to prevent accidental repository lock outs.

"registryId": The Amazon Web Services account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

Main.Ecr.start_image_scan — Function

start_image_scan(image_id, repository_name)
start_image_scan(image_id, repository_name, params::Dict{String,<:Any})

Starts an image vulnerability scan. An image scan can only be started once per 24 hours on an individual image. This limit includes if an image was scanned on initial push. For more information, see Image scanning in the Amazon Elastic Container Registry User Guide.

Arguments

image_id:

repository_name: The name of the repository that contains the images to scan.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"registryId": The Amazon Web Services account ID associated with the registry that contains the repository in which to start an image scan request. If you do not specify a registry, the default registry is assumed.

Main.Ecr.start_lifecycle_policy_preview — Function

start_lifecycle_policy_preview(repository_name)
start_lifecycle_policy_preview(repository_name, params::Dict{String,<:Any})

Starts a preview of a lifecycle policy for the specified repository. This allows you to see the results before associating the lifecycle policy with the repository.

Arguments

repository_name: The name of the repository to be evaluated.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"lifecyclePolicyText": The policy to be evaluated against. If you do not specify a policy, the current policy for the repository is used.

"registryId": The Amazon Web Services account ID associated with the registry that contains the repository. If you do not specify a registry, the default registry is assumed.

Main.Ecr.tag_resource — Function

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Adds specified tags to a resource with the specified ARN. Existing tags on a resource are not changed if they are not specified in the request parameters.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the the resource to which to add tags. Currently, the only supported resource is an Amazon ECR repository.

tags: The tags to add to the resource. A tag is an array of key-value pairs. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.

Main.Ecr.untag_resource — Function

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Deletes specified tags from a resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource from which to remove tags. Currently, the only supported resource is an Amazon ECR repository.

tag_keys: The keys of the tags to be removed.

Main.Ecr.update_pull_through_cache_rule — Function

update_pull_through_cache_rule(credential_arn, ecr_repository_prefix)
update_pull_through_cache_rule(credential_arn, ecr_repository_prefix, params::Dict{String,<:Any})

Updates an existing pull through cache rule.

Arguments

credential_arn: The Amazon Resource Name (ARN) of the Amazon Web Services Secrets Manager secret that identifies the credentials to authenticate to the upstream registry.

ecr_repository_prefix: The repository name prefix to use when caching images from the source registry.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"registryId": The Amazon Web Services account ID associated with the registry associated with the pull through cache rule. If you do not specify a registry, the default registry is assumed.

Main.Ecr.update_repository_creation_template — Function

update_repository_creation_template(prefix)
update_repository_creation_template(prefix, params::Dict{String,<:Any})

Updates an existing repository creation template.

Arguments

prefix: The repository namespace prefix that matches an existing repository creation template in the registry. All repositories created using this namespace prefix will have the settings defined in this template applied. For example, a prefix of prod would apply to all repositories beginning with prod/. This includes a repository named prod/team1 as well as a repository named prod/repository1. To apply a template to all repositories in your registry that don't have an associated creation template, you can use ROOT as the prefix.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"appliedFor": Updates the list of enumerable strings representing the Amazon ECR repository creation scenarios that this template will apply towards. The two supported scenarios are PULLTHROUGHCACHE and REPLICATION

"customRoleArn": The ARN of the role to be assumed by Amazon ECR. This role must be in the same account as the registry that you are configuring. Amazon ECR will assume your supplied role when the customRoleArn is specified. When this field isn't specified, Amazon ECR will use the service-linked role for the repository creation template.

"description": A description for the repository creation template.

"encryptionConfiguration":

"imageTagMutability": Updates the tag mutability setting for the repository. If this parameter is omitted, the default setting of MUTABLE will be used which will allow image tags to be overwritten. If IMMUTABLE is specified, all image tags within the repository will be immutable which will prevent them from being overwritten.

"lifecyclePolicy": Updates the lifecycle policy associated with the specified repository creation template.

"repositoryPolicy": Updates the repository policy created using the template. A repository policy is a permissions policy associated with a repository to control access permissions.

"resourceTags": The metadata to apply to the repository to help you categorize and organize. Each tag consists of a key and an optional value, both of which you define. Tag keys can have a maximum character length of 128 characters, and tag values can have a maximum length of 256 characters.

Main.Ecr.upload_layer_part — Function

upload_layer_part(layer_part_blob, part_first_byte, part_last_byte, repository_name, upload_id)
upload_layer_part(layer_part_blob, part_first_byte, part_last_byte, repository_name, upload_id, params::Dict{String,<:Any})

Uploads an image layer part to Amazon ECR. When an image is pushed, each new image layer is uploaded in parts. The maximum size of each image layer part can be 20971520 bytes (or about 20MB). The UploadLayerPart API is called once per each new image layer part. This operation is used by the Amazon ECR proxy and is not generally used by customers for pulling and pushing images. In most cases, you should use the docker CLI to pull, tag, and push images.

Arguments

layer_part_blob: The base64-encoded layer part payload.

part_first_byte: The position of the first byte of the layer part witin the overall image layer.

part_last_byte: The position of the last byte of the layer part within the overall image layer.

repository_name: The name of the repository to which you are uploading layer parts.

upload_id: The upload ID from a previous InitiateLayerUpload operation to associate with the layer part upload.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"registryId": The Amazon Web Services account ID associated with the registry to which you are uploading layer parts. If you do not specify a registry, the default registry is assumed.

Main.Ecr.validate_pull_through_cache_rule — Function

validate_pull_through_cache_rule(ecr_repository_prefix)
validate_pull_through_cache_rule(ecr_repository_prefix, params::Dict{String,<:Any})

Validates an existing pull through cache rule for an upstream registry that requires authentication. This will retrieve the contents of the Amazon Web Services Secrets Manager secret, verify the syntax, and then validate that authentication to the upstream registry is successful.

Arguments

ecr_repository_prefix: The repository name prefix associated with the pull through cache rule.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"registryId": The registry ID associated with the pull through cache rule. If you do not specify a registry, the default registry is assumed.