Emr Serverless

This page documents function available when using the Emr_Serverless module, created with @service Emr_Serverless.

Index

Main.Emr_Serverless.cancel_job_run

Main.Emr_Serverless.create_application

Main.Emr_Serverless.delete_application

Main.Emr_Serverless.get_application

Main.Emr_Serverless.get_dashboard_for_job_run

Main.Emr_Serverless.get_job_run

Main.Emr_Serverless.list_applications

Main.Emr_Serverless.list_job_runs

Main.Emr_Serverless.list_tags_for_resource

Main.Emr_Serverless.start_application

Main.Emr_Serverless.start_job_run

Main.Emr_Serverless.stop_application

Main.Emr_Serverless.tag_resource

Main.Emr_Serverless.untag_resource

Main.Emr_Serverless.update_application

Documentation

Main.Emr_Serverless.cancel_job_run — Method

cancel_job_run(application_id, job_run_id)
cancel_job_run(application_id, job_run_id, params::Dict{String,<:Any})

Cancels a job run.

Arguments

application_id: The ID of the application on which the job run will be canceled.

job_run_id: The ID of the job run to cancel.

Main.Emr_Serverless.create_application — Method

create_application(client_token, release_label, type)
create_application(client_token, release_label, type, params::Dict{String,<:Any})

Creates an application.

Arguments

client_token: The client idempotency token of the application to create. Its value must be unique for each request.

release_label: The EMR release associated with the application.

type: The type of application you want to start, such as Spark or Hive.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"architecture": The CPU architecture of an application.

"autoStartConfiguration": The configuration for an application to automatically start on job submission.

"autoStopConfiguration": The configuration for an application to automatically stop after a certain amount of time being idle.

"imageConfiguration": The image configuration for all worker types. You can either set this parameter or imageConfiguration for each worker type in workerTypeSpecifications.

"initialCapacity": The capacity to initialize when the application is created.

"maximumCapacity": The maximum capacity to allocate when the application is created. This is cumulative across all workers at any given point in time, not just when an application is created. No new resources will be created once any one of the defined limits is hit.

"name": The name of the application.

"networkConfiguration": The network configuration for customer VPC connectivity.

"tags": The tags assigned to the application.

"workerTypeSpecifications": The key-value pairs that specify worker type to WorkerTypeSpecificationInput. This parameter must contain all valid worker types for a Spark or Hive application. Valid worker types include Driver and Executor for Spark applications and HiveDriver and TezTask for Hive applications. You can either set image details in this parameter for each worker type, or in imageConfiguration for all worker types.

Main.Emr_Serverless.delete_application — Method

delete_application(application_id)
delete_application(application_id, params::Dict{String,<:Any})

Deletes an application. An application has to be in a stopped or created state in order to be deleted.

Arguments

application_id: The ID of the application that will be deleted.

Main.Emr_Serverless.get_application — Method

get_application(application_id)
get_application(application_id, params::Dict{String,<:Any})

Displays detailed information about a specified application.

Arguments

application_id: The ID of the application that will be described.

Main.Emr_Serverless.get_dashboard_for_job_run — Method

get_dashboard_for_job_run(application_id, job_run_id)
get_dashboard_for_job_run(application_id, job_run_id, params::Dict{String,<:Any})

Returns a URL to access the job run dashboard. The generated URL is valid for one hour, after which you must invoke the API again to generate a new URL.

Arguments

application_id: The ID of the application.

job_run_id: The ID of the job run.

Main.Emr_Serverless.get_job_run — Method

get_job_run(application_id, job_run_id)
get_job_run(application_id, job_run_id, params::Dict{String,<:Any})

Displays detailed information about a job run.

Arguments

application_id: The ID of the application on which the job run is submitted.

job_run_id: The ID of the job run.

Main.Emr_Serverless.list_applications — Method

list_applications()
list_applications(params::Dict{String,<:Any})

Lists applications based on a set of parameters.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of applications that can be listed.

"nextToken": The token for the next set of application results.

"states": An optional filter for application states. Note that if this filter contains multiple states, the resulting list will be grouped by the state.

Main.Emr_Serverless.list_job_runs — Method

list_job_runs(application_id)
list_job_runs(application_id, params::Dict{String,<:Any})

Lists job runs based on a set of parameters.

Arguments

application_id: The ID of the application for which to list the job run.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"createdAtAfter": The lower bound of the option to filter by creation date and time.

"createdAtBefore": The upper bound of the option to filter by creation date and time.

"maxResults": The maximum number of job runs that can be listed.

"nextToken": The token for the next set of job run results.

"states": An optional filter for job run states. Note that if this filter contains multiple states, the resulting list will be grouped by the state.

Main.Emr_Serverless.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists the tags assigned to the resources.

Arguments

resource_arn: The Amazon Resource Name (ARN) that identifies the resource to list the tags for. Currently, the supported resources are Amazon EMR Serverless applications and job runs.

Main.Emr_Serverless.start_application — Method

start_application(application_id)
start_application(application_id, params::Dict{String,<:Any})

Starts a specified application and initializes initial capacity if configured.

Arguments

application_id: The ID of the application to start.

Main.Emr_Serverless.start_job_run — Method

start_job_run(application_id, client_token, execution_role_arn)
start_job_run(application_id, client_token, execution_role_arn, params::Dict{String,<:Any})

Starts a job run.

Arguments

application_id: The ID of the application on which to run the job.

client_token: The client idempotency token of the job run to start. Its value must be unique for each request.

execution_role_arn: The execution role ARN for the job run.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"configurationOverrides": The configuration overrides for the job run.

"executionTimeoutMinutes": The maximum duration for the job run to run. If the job run runs beyond this duration, it will be automatically cancelled.

"jobDriver": The job driver for the job run.

"name": The optional job run name. This doesn't have to be unique.

"tags": The tags assigned to the job run.

Main.Emr_Serverless.stop_application — Method

stop_application(application_id)
stop_application(application_id, params::Dict{String,<:Any})

Stops a specified application and releases initial capacity if configured. All scheduled and running jobs must be completed or cancelled before stopping an application.

Arguments

application_id: The ID of the application to stop.

Main.Emr_Serverless.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Assigns tags to resources. A tag is a label that you assign to an Amazon Web Services resource. Each tag consists of a key and an optional value, both of which you define. Tags enable you to categorize your Amazon Web Services resources by attributes such as purpose, owner, or environment. When you have many resources of the same type, you can quickly identify a specific resource based on the tags you've assigned to it.

Arguments

resource_arn: The Amazon Resource Name (ARN) that identifies the resource to list the tags for. Currently, the supported resources are Amazon EMR Serverless applications and job runs.

tags: The tags to add to the resource. A tag is an array of key-value pairs.

Main.Emr_Serverless.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes tags from resources.

Arguments

resource_arn: The Amazon Resource Name (ARN) that identifies the resource to list the tags for. Currently, the supported resources are Amazon EMR Serverless applications and job runs.

tag_keys: The keys of the tags to be removed.

Main.Emr_Serverless.update_application — Method

update_application(application_id, client_token)
update_application(application_id, client_token, params::Dict{String,<:Any})

Updates a specified application. An application has to be in a stopped or created state in order to be updated.

Arguments

application_id: The ID of the application to update.

client_token: The client idempotency token of the application to update. Its value must be unique for each request.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"architecture": The CPU architecture of an application.

"autoStartConfiguration": The configuration for an application to automatically start on job submission.

"autoStopConfiguration": The configuration for an application to automatically stop after a certain amount of time being idle.

"imageConfiguration": The image configuration to be used for all worker types. You can either set this parameter or imageConfiguration for each worker type in WorkerTypeSpecificationInput.

"initialCapacity": The capacity to initialize when the application is updated.

"maximumCapacity": The maximum capacity to allocate when the application is updated. This is cumulative across all workers at any given point in time during the lifespan of the application. No new resources will be created once any one of the defined limits is hit.

"networkConfiguration":

"workerTypeSpecifications": The key-value pairs that specify worker type to WorkerTypeSpecificationInput. This parameter must contain all valid worker types for a Spark or Hive application. Valid worker types include Driver and Executor for Spark applications and HiveDriver and TezTask for Hive applications. You can either set image details in this parameter for each worker type, or in imageConfiguration for all worker types.