Entityresolution

This page documents function available when using the Entityresolution module, created with @service Entityresolution.

Index

Main.Entityresolution.add_policy_statement

Main.Entityresolution.batch_delete_unique_id

Main.Entityresolution.create_id_mapping_workflow

Main.Entityresolution.create_id_namespace

Main.Entityresolution.create_matching_workflow

Main.Entityresolution.create_schema_mapping

Main.Entityresolution.delete_id_mapping_workflow

Main.Entityresolution.delete_id_namespace

Main.Entityresolution.delete_matching_workflow

Main.Entityresolution.delete_policy_statement

Main.Entityresolution.delete_schema_mapping

Main.Entityresolution.get_id_mapping_job

Main.Entityresolution.get_id_mapping_workflow

Main.Entityresolution.get_id_namespace

Main.Entityresolution.get_match_id

Main.Entityresolution.get_matching_job

Main.Entityresolution.get_matching_workflow

Main.Entityresolution.get_policy

Main.Entityresolution.get_provider_service

Main.Entityresolution.get_schema_mapping

Main.Entityresolution.list_id_mapping_jobs

Main.Entityresolution.list_id_mapping_workflows

Main.Entityresolution.list_id_namespaces

Main.Entityresolution.list_matching_jobs

Main.Entityresolution.list_matching_workflows

Main.Entityresolution.list_provider_services

Main.Entityresolution.list_schema_mappings

Main.Entityresolution.list_tags_for_resource

Main.Entityresolution.put_policy

Main.Entityresolution.start_id_mapping_job

Main.Entityresolution.start_matching_job

Main.Entityresolution.tag_resource

Main.Entityresolution.untag_resource

Main.Entityresolution.update_id_mapping_workflow

Main.Entityresolution.update_id_namespace

Main.Entityresolution.update_matching_workflow

Main.Entityresolution.update_schema_mapping

Documentation

Main.Entityresolution.add_policy_statement — Method

add_policy_statement(action, arn, effect, principal, statement_id)
add_policy_statement(action, arn, effect, principal, statement_id, params::Dict{String,<:Any})

Adds a policy statement object. To retrieve a list of existing policy statements, use the GetPolicy API.

Arguments

action: The action that the principal can use on the resource. For example, entityresolution:GetIdMappingJob, entityresolution:GetMatchingJob.

arn: The Amazon Resource Name (ARN) of the resource that will be accessed by the principal.

effect: Determines whether the permissions specified in the policy are to be allowed (Allow) or denied (Deny).

principal: The Amazon Web Services service or Amazon Web Services account that can access the resource defined as ARN.

statement_id: A statement identifier that differentiates the statement from others in the same policy.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"condition": A set of condition keys that you can use in key policies.

Main.Entityresolution.batch_delete_unique_id — Method

batch_delete_unique_id(unique_ids, workflow_name)
batch_delete_unique_id(unique_ids, workflow_name, params::Dict{String,<:Any})

Deletes multiple unique IDs in a matching workflow.

Arguments

unique_ids: The unique IDs to delete.

workflow_name: The name of the workflow.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"inputSource": The input source for the batch delete unique ID operation.

Main.Entityresolution.create_id_mapping_workflow — Method

create_id_mapping_workflow(id_mapping_techniques, input_source_config, role_arn, workflow_name)
create_id_mapping_workflow(id_mapping_techniques, input_source_config, role_arn, workflow_name, params::Dict{String,<:Any})

Creates an IdMappingWorkflow object which stores the configuration of the data processing job to be run. Each IdMappingWorkflow must have a unique workflow name. To modify an existing workflow, use the UpdateIdMappingWorkflow API.

Arguments

id_mapping_techniques: An object which defines the idMappingType and the providerProperties.

input_source_config: A list of InputSource objects, which have the fields InputSourceARN and SchemaName.

role_arn: The Amazon Resource Name (ARN) of the IAM role. Entity Resolution assumes this role to create resources on your behalf as part of workflow execution.

workflow_name: The name of the workflow. There can't be multiple IdMappingWorkflows with the same name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A description of the workflow.

"outputSourceConfig": A list of IdMappingWorkflowOutputSource objects, each of which contains fields OutputS3Path and Output.

"tags": The tags used to organize, track, or control access for this resource.

Main.Entityresolution.create_id_namespace — Method

create_id_namespace(id_namespace_name, type)
create_id_namespace(id_namespace_name, type, params::Dict{String,<:Any})

Creates an ID namespace object which will help customers provide metadata explaining their dataset and how to use it. Each ID namespace must have a unique name. To modify an existing ID namespace, use the UpdateIdNamespace API.

Arguments

id_namespace_name: The name of the ID namespace.

type: The type of ID namespace. There are two types: SOURCE and TARGET. The SOURCE contains configurations for sourceId data that will be processed in an ID mapping workflow. The TARGET contains a configuration of targetId to which all sourceIds will resolve to.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The description of the ID namespace.

"idMappingWorkflowProperties": Determines the properties of IdMappingWorflow where this IdNamespace can be used as a Source or a Target.

"inputSourceConfig": A list of InputSource objects, which have the fields InputSourceARN and SchemaName.

"roleArn": The Amazon Resource Name (ARN) of the IAM role. Entity Resolution assumes this role to access the resources defined in this IdNamespace on your behalf as part of the workflow run.

"tags": The tags used to organize, track, or control access for this resource.

Main.Entityresolution.create_matching_workflow — Method

create_matching_workflow(input_source_config, output_source_config, resolution_techniques, role_arn, workflow_name)
create_matching_workflow(input_source_config, output_source_config, resolution_techniques, role_arn, workflow_name, params::Dict{String,<:Any})

Creates a MatchingWorkflow object which stores the configuration of the data processing job to be run. It is important to note that there should not be a pre-existing MatchingWorkflow with the same name. To modify an existing workflow, utilize the UpdateMatchingWorkflow API.

Arguments

input_source_config: A list of InputSource objects, which have the fields InputSourceARN and SchemaName.

output_source_config: A list of OutputSource objects, each of which contains fields OutputS3Path, ApplyNormalization, and Output.

resolution_techniques: An object which defines the resolutionType and the ruleBasedProperties.

role_arn: The Amazon Resource Name (ARN) of the IAM role. Entity Resolution assumes this role to create resources on your behalf as part of workflow execution.

workflow_name: The name of the workflow. There can't be multiple MatchingWorkflows with the same name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A description of the workflow.

"incrementalRunConfig": An object which defines an incremental run type and has only incrementalRunType as a field.

"tags": The tags used to organize, track, or control access for this resource.

Main.Entityresolution.create_schema_mapping — Method

create_schema_mapping(mapped_input_fields, schema_name)
create_schema_mapping(mapped_input_fields, schema_name, params::Dict{String,<:Any})

Creates a schema mapping, which defines the schema of the input customer records table. The SchemaMapping also provides Entity Resolution with some metadata about the table, such as the attribute types of the columns and which columns to match on.

Arguments

mapped_input_fields: A list of MappedInputFields. Each MappedInputField corresponds to a column the source data table, and contains column name plus additional information that Entity Resolution uses for matching.

schema_name: The name of the schema. There can't be multiple SchemaMappings with the same name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A description of the schema.

"tags": The tags used to organize, track, or control access for this resource.

Main.Entityresolution.delete_id_mapping_workflow — Method

delete_id_mapping_workflow(workflow_name)
delete_id_mapping_workflow(workflow_name, params::Dict{String,<:Any})

Deletes the IdMappingWorkflow with a given name. This operation will succeed even if a workflow with the given name does not exist.

Arguments

workflow_name: The name of the workflow to be deleted.

Main.Entityresolution.delete_id_namespace — Method

delete_id_namespace(id_namespace_name)
delete_id_namespace(id_namespace_name, params::Dict{String,<:Any})

Deletes the IdNamespace with a given name.

Arguments

id_namespace_name: The name of the ID namespace.

Main.Entityresolution.delete_matching_workflow — Method

delete_matching_workflow(workflow_name)
delete_matching_workflow(workflow_name, params::Dict{String,<:Any})

Deletes the MatchingWorkflow with a given name. This operation will succeed even if a workflow with the given name does not exist.

Arguments

workflow_name: The name of the workflow to be retrieved.

Main.Entityresolution.delete_policy_statement — Method

delete_policy_statement(arn, statement_id)
delete_policy_statement(arn, statement_id, params::Dict{String,<:Any})

Deletes the policy statement.

Arguments

arn: The ARN of the resource for which the policy need to be deleted.

statement_id: A statement identifier that differentiates the statement from others in the same policy.

Main.Entityresolution.delete_schema_mapping — Method

delete_schema_mapping(schema_name)
delete_schema_mapping(schema_name, params::Dict{String,<:Any})

Deletes the SchemaMapping with a given name. This operation will succeed even if a schema with the given name does not exist. This operation will fail if there is a MatchingWorkflow object that references the SchemaMapping in the workflow's InputSourceConfig.

Arguments

schema_name: The name of the schema to delete.

Main.Entityresolution.get_id_mapping_job — Method

get_id_mapping_job(job_id, workflow_name)
get_id_mapping_job(job_id, workflow_name, params::Dict{String,<:Any})

Gets the status, metrics, and errors (if there are any) that are associated with a job.

Arguments

job_id: The ID of the job.

workflow_name: The name of the workflow.

Main.Entityresolution.get_id_mapping_workflow — Method

get_id_mapping_workflow(workflow_name)
get_id_mapping_workflow(workflow_name, params::Dict{String,<:Any})

Returns the IdMappingWorkflow with a given name, if it exists.

Arguments

workflow_name: The name of the workflow.

Main.Entityresolution.get_id_namespace — Method

get_id_namespace(id_namespace_name)
get_id_namespace(id_namespace_name, params::Dict{String,<:Any})

Returns the IdNamespace with a given name, if it exists.

Arguments

id_namespace_name: The name of the ID namespace.

Main.Entityresolution.get_match_id — Method

get_match_id(record, workflow_name)
get_match_id(record, workflow_name, params::Dict{String,<:Any})

Returns the corresponding Match ID of a customer record if the record has been processed.

Arguments

record: The record to fetch the Match ID for.

workflow_name: The name of the workflow.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"applyNormalization": Normalizes the attributes defined in the schema in the input data. For example, if an attribute has an AttributeType of PHONE_NUMBER, and the data in the input table is in a format of 1234567890, Entity Resolution will normalize this field in the output to (123)-456-7890.

Main.Entityresolution.get_matching_job — Method

get_matching_job(job_id, workflow_name)
get_matching_job(job_id, workflow_name, params::Dict{String,<:Any})

Gets the status, metrics, and errors (if there are any) that are associated with a job.

Arguments

job_id: The ID of the job.

workflow_name: The name of the workflow.

Main.Entityresolution.get_matching_workflow — Method

get_matching_workflow(workflow_name)
get_matching_workflow(workflow_name, params::Dict{String,<:Any})

Returns the MatchingWorkflow with a given name, if it exists.

Arguments

workflow_name: The name of the workflow.

Main.Entityresolution.get_policy — Method

get_policy(arn)
get_policy(arn, params::Dict{String,<:Any})

Returns the resource-based policy.

Arguments

arn: The Amazon Resource Name (ARN) of the resource for which the policy need to be returned.

Main.Entityresolution.get_provider_service — Method

get_provider_service(provider_name, provider_service_name)
get_provider_service(provider_name, provider_service_name, params::Dict{String,<:Any})

Returns the ProviderService of a given name.

Arguments

provider_name: The name of the provider. This name is typically the company name.

provider_service_name: The ARN (Amazon Resource Name) of the product that the provider service provides.

Main.Entityresolution.get_schema_mapping — Method

get_schema_mapping(schema_name)
get_schema_mapping(schema_name, params::Dict{String,<:Any})

Returns the SchemaMapping of a given name.

Arguments

schema_name: The name of the schema to be retrieved.

Main.Entityresolution.list_id_mapping_jobs — Method

list_id_mapping_jobs(workflow_name)
list_id_mapping_jobs(workflow_name, params::Dict{String,<:Any})

Lists all ID mapping jobs for a given workflow.

Arguments

workflow_name: The name of the workflow to be retrieved.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of objects returned per page.

"nextToken": The pagination token from the previous API call.

Main.Entityresolution.list_id_mapping_workflows — Method

list_id_mapping_workflows()
list_id_mapping_workflows(params::Dict{String,<:Any})

Returns a list of all the IdMappingWorkflows that have been created for an Amazon Web Services account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of objects returned per page.

"nextToken": The pagination token from the previous API call.

Main.Entityresolution.list_id_namespaces — Method

list_id_namespaces()
list_id_namespaces(params::Dict{String,<:Any})

Returns a list of all ID namespaces.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of IdNamespace objects returned per page.

"nextToken": The pagination token from the previous API call.

Main.Entityresolution.list_matching_jobs — Method

list_matching_jobs(workflow_name)
list_matching_jobs(workflow_name, params::Dict{String,<:Any})

Lists all jobs for a given workflow.

Arguments

workflow_name: The name of the workflow to be retrieved.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of objects returned per page.

"nextToken": The pagination token from the previous API call.

Main.Entityresolution.list_matching_workflows — Method

list_matching_workflows()
list_matching_workflows(params::Dict{String,<:Any})

Returns a list of all the MatchingWorkflows that have been created for an Amazon Web Services account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of objects returned per page.

"nextToken": The pagination token from the previous API call.

Main.Entityresolution.list_provider_services — Method

list_provider_services()
list_provider_services(params::Dict{String,<:Any})

Returns a list of all the ProviderServices that are available in this Amazon Web Services Region.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of objects returned per page.

"nextToken": The pagination token from the previous API call.

"providerName": The name of the provider. This name is typically the company name.

Main.Entityresolution.list_schema_mappings — Method

list_schema_mappings()
list_schema_mappings(params::Dict{String,<:Any})

Returns a list of all the SchemaMappings that have been created for an Amazon Web Services account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of objects returned per page.

"nextToken": The pagination token from the previous API call.

Main.Entityresolution.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Displays the tags associated with an Entity Resolution resource. In Entity Resolution, SchemaMapping, and MatchingWorkflow can be tagged.

Arguments

resource_arn: The ARN of the resource for which you want to view tags.

Main.Entityresolution.put_policy — Method

put_policy(arn, policy)
put_policy(arn, policy, params::Dict{String,<:Any})

Updates the resource-based policy.

Arguments

arn: The Amazon Resource Name (ARN) of the resource for which the policy needs to be updated.

policy: The resource-based policy.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"token": A unique identifier for the current revision of the policy.

Main.Entityresolution.start_id_mapping_job — Method

start_id_mapping_job(workflow_name)
start_id_mapping_job(workflow_name, params::Dict{String,<:Any})

Starts the IdMappingJob of a workflow. The workflow must have previously been created using the CreateIdMappingWorkflow endpoint.

Arguments

workflow_name: The name of the ID mapping job to be retrieved.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"outputSourceConfig": A list of OutputSource objects.

Main.Entityresolution.start_matching_job — Method

start_matching_job(workflow_name)
start_matching_job(workflow_name, params::Dict{String,<:Any})

Starts the MatchingJob of a workflow. The workflow must have previously been created using the CreateMatchingWorkflow endpoint.

Arguments

workflow_name: The name of the matching job to be retrieved.

Main.Entityresolution.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Assigns one or more tags (key-value pairs) to the specified Entity Resolution resource. Tags can help you organize and categorize your resources. You can also use them to scope user permissions by granting a user permission to access or change only resources with certain tag values. In Entity Resolution, SchemaMapping and MatchingWorkflow can be tagged. Tags don't have any semantic meaning to Amazon Web Services and are interpreted strictly as strings of characters. You can use the TagResource action with a resource that already has tags. If you specify a new tag key, this tag is appended to the list of tags associated with the resource. If you specify a tag key that is already associated with the resource, the new tag value that you specify replaces the previous value for that tag.

Arguments

resource_arn: The ARN of the resource for which you want to view tags.

tags: The tags used to organize, track, or control access for this resource.

Main.Entityresolution.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes one or more tags from the specified Entity Resolution resource. In Entity Resolution, SchemaMapping, and MatchingWorkflow can be tagged.

Arguments

resource_arn: The ARN of the resource for which you want to untag.

tag_keys: The list of tag keys to remove from the resource.

Main.Entityresolution.update_id_mapping_workflow — Method

update_id_mapping_workflow(id_mapping_techniques, input_source_config, role_arn, workflow_name)
update_id_mapping_workflow(id_mapping_techniques, input_source_config, role_arn, workflow_name, params::Dict{String,<:Any})

Updates an existing IdMappingWorkflow. This method is identical to CreateIdMappingWorkflow, except it uses an HTTP PUT request instead of a POST request, and the IdMappingWorkflow must already exist for the method to succeed.

Arguments

id_mapping_techniques: An object which defines the idMappingType and the providerProperties.

input_source_config: A list of InputSource objects, which have the fields InputSourceARN and SchemaName.

role_arn: The Amazon Resource Name (ARN) of the IAM role. Entity Resolution assumes this role to access Amazon Web Services resources on your behalf.

workflow_name: The name of the workflow.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A description of the workflow.

"outputSourceConfig": A list of OutputSource objects, each of which contains fields OutputS3Path and KMSArn.

Main.Entityresolution.update_id_namespace — Method

update_id_namespace(id_namespace_name)
update_id_namespace(id_namespace_name, params::Dict{String,<:Any})

Updates an existing ID namespace.

Arguments

id_namespace_name: The name of the ID namespace.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The description of the ID namespace.

"idMappingWorkflowProperties": Determines the properties of IdMappingWorkflow where this IdNamespace can be used as a Source or a Target.

"inputSourceConfig": A list of InputSource objects, which have the fields InputSourceARN and SchemaName.

"roleArn": The Amazon Resource Name (ARN) of the IAM role. Entity Resolution assumes this role to access the resources defined in this IdNamespace on your behalf as part of a workflow run.

Main.Entityresolution.update_matching_workflow — Method

update_matching_workflow(input_source_config, output_source_config, resolution_techniques, role_arn, workflow_name)
update_matching_workflow(input_source_config, output_source_config, resolution_techniques, role_arn, workflow_name, params::Dict{String,<:Any})

Updates an existing MatchingWorkflow. This method is identical to CreateMatchingWorkflow, except it uses an HTTP PUT request instead of a POST request, and the MatchingWorkflow must already exist for the method to succeed.

Arguments

input_source_config: A list of InputSource objects, which have the fields InputSourceARN and SchemaName.

output_source_config: A list of OutputSource objects, each of which contains fields OutputS3Path, ApplyNormalization, and Output.

resolution_techniques: An object which defines the resolutionType and the ruleBasedProperties.

role_arn: The Amazon Resource Name (ARN) of the IAM role. Entity Resolution assumes this role to create resources on your behalf as part of workflow execution.

workflow_name: The name of the workflow to be retrieved.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A description of the workflow.

"incrementalRunConfig": An object which defines an incremental run type and has only incrementalRunType as a field.

Main.Entityresolution.update_schema_mapping — Method

update_schema_mapping(mapped_input_fields, schema_name)
update_schema_mapping(mapped_input_fields, schema_name, params::Dict{String,<:Any})

Updates a schema mapping. A schema is immutable if it is being used by a workflow. Therefore, you can't update a schema mapping if it's associated with a workflow.

Arguments

mapped_input_fields: A list of MappedInputFields. Each MappedInputField corresponds to a column the source data table, and contains column name plus additional information that Entity Resolution uses for matching.

schema_name: The name of the schema. There can't be multiple SchemaMappings with the same name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A description of the schema.