Evs

This page documents function available when using the Evs module, created with @service Evs.

Index

Main.Evs.associate_eip_to_vlan

Main.Evs.create_environment

Main.Evs.create_environment_host

Main.Evs.delete_environment

Main.Evs.delete_environment_host

Main.Evs.disassociate_eip_from_vlan

Main.Evs.get_environment

Main.Evs.get_versions

Main.Evs.list_environment_hosts

Main.Evs.list_environment_vlans

Main.Evs.list_environments

Main.Evs.list_tags_for_resource

Main.Evs.tag_resource

Main.Evs.untag_resource

Documentation

Main.Evs.associate_eip_to_vlan — Function

associate_eip_to_vlan(allocation_id, environment_id, vlan_name)
associate_eip_to_vlan(allocation_id, environment_id, vlan_name, params::Dict{String,<:Any})

Associates an Elastic IP address with a public HCX VLAN. This operation is only allowed for public HCX VLANs at this time.

Arguments

allocation_id: The Elastic IP address allocation ID.

environment_id: A unique ID for the environment containing the VLAN that the Elastic IP address associates with.

vlan_name: The name of the VLAN. hcx is the only accepted VLAN name at this time.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": !!! note This parameter is not used in Amazon EVS currently. If you supply input for this parameter, it will have no effect.

A unique, case-sensitive identifier that you provide to ensure the idempotency of the environment creation request. If you do not specify a client token, a randomly generated token is used for the request to ensure idempotency.

Main.Evs.create_environment — Function

create_environment(connectivity_info, hosts, initial_vlans, license_info, service_access_subnet_id, site_id, terms_accepted, vcf_hostnames, vcf_version, vpc_id)
create_environment(connectivity_info, hosts, initial_vlans, license_info, service_access_subnet_id, site_id, terms_accepted, vcf_hostnames, vcf_version, vpc_id, params::Dict{String,<:Any})

Creates an Amazon EVS environment that runs VCF software, such as SDDC Manager, NSX Manager, and vCenter Server.

During environment creation, Amazon EVS performs validations on DNS settings, provisions VLAN subnets and hosts, and deploys the supplied version of VCF.

It can take several hours to create an environment. After the deployment completes, you can configure VCF in the vSphere user interface according to your needs.

Important

When creating a new environment, the default ESX version for the selected VCF version will be used, you cannot choose a specific ESX version in CreateEnvironment action. When a host has been added with a specific ESX version, it can only be upgraded using vCenter Lifecycle Manager.

Note

You cannot use the dedicatedHostId and placementGroupId parameters together in the same CreateEnvironment action. This results in a ValidationException response.

Arguments

connectivity_info: The connectivity configuration for the environment. Amazon EVS requires that you specify two route server peer IDs. During environment creation, the route server endpoints peer with the NSX edges over the NSX uplink subnet, providing BGP- based dynamic routing for overlay networks.

hosts: The ESX hosts to add to the environment. Amazon EVS requires that you provide details for a minimum of 4 hosts during environment creation.

For each host, you must provide the desired hostname, EC2 SSH keypair name, and EC2 instance type. Optionally, you can also provide a partition or cluster placement group to use, or use Amazon EC2 Dedicated Hosts.

initial_vlans: The initial VLAN subnets for the Amazon EVS environment.

Note

For each Amazon EVS VLAN subnet, you must specify a non-overlapping CIDR block. Amazon EVS VLAN subnets have a minimum CIDR block size of /28 and a maximum size of /24.

license_info: The license information that Amazon EVS requires to create an environment. Amazon EVS requires two license keys: a VCF solution key and a vSAN license key. The VCF solution key must cover a minimum of 256 cores. The vSAN license key must provide at least 110 TiB of vSAN capacity.

VCF licenses can be used for only one Amazon EVS environment. Amazon EVS does not support reuse of VCF licenses for multiple environments.

VCF license information can be retrieved from the Broadcom portal.

service_access_subnet_id: The subnet that is used to establish connectivity between the Amazon EVS control plane and VPC. Amazon EVS uses this subnet to validate mandatory DNS records for your VCF appliances and hosts and create the environment.

site_id: The Broadcom Site ID that is allocated to you as part of your electronic software delivery. This ID allows customer access to the Broadcom portal, and is provided to you by Broadcom at the close of your software contract or contract renewal. Amazon EVS uses the Broadcom Site ID that you provide to meet Broadcom VCF license usage reporting requirements for Amazon EVS.

terms_accepted: Customer confirmation that the customer has purchased and will continue to maintain the required number of VCF software licenses to cover all physical processor cores in the Amazon EVS environment. Information about your VCF software in Amazon EVS will be shared with Broadcom to verify license compliance. Amazon EVS does not validate license keys. To validate license keys, visit the Broadcom support portal.

vcf_hostnames: The DNS hostnames for the virtual machines that host the VCF management appliances. Amazon EVS requires that you provide DNS hostnames for the following appliances: vCenter, NSX Manager, SDDC Manager, and Cloud Builder.

vcf_version: The VCF version to use for the environment.

vpc_id: A unique ID for the VPC that the environment is deployed inside.

Amazon EVS requires that all VPC subnets exist in a single Availability Zone in a Region where the service is available.

The VPC that you specify must have a valid DHCP option set with domain name, at least two DNS servers, and an NTP server. These settings are used to configure your VCF appliances and hosts. The VPC cannot be used with any other deployed Amazon EVS environment. Amazon EVS does not provide multi-VPC support for environments at this time.

Amazon EVS does not support the following Amazon Web Services networking options for NSX overlay connectivity: cross-Region VPC peering, Amazon S3 gateway endpoints, or Amazon Web Services Direct Connect virtual private gateway associations.

Note

Ensure that you specify a VPC that is adequately sized to accommodate the Amazon EVS subnets.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": !!! note This parameter is not used in Amazon EVS currently. If you supply input for this parameter, it will have no effect.

A unique, case-sensitive identifier that you provide to ensure the idempotency of the environment creation request. If you do not specify a client token, a randomly generated token is used for the request to ensure idempotency.

"environmentName": The name to give to your environment. The name can contain only alphanumeric characters (case-sensitive), hyphens, and underscores. It must start with an alphanumeric character, and can't be longer than 100 characters. The name must be unique within the Amazon Web Services Region and Amazon Web Services account that you're creating the environment in.

"kmsKeyId": A unique ID for the customer-managed KMS key that is used to encrypt the VCF credential pairs for SDDC Manager, NSX Manager, and vCenter appliances. These credentials are stored in Amazon Web Services Secrets Manager.

"serviceAccessSecurityGroups": The security group that controls communication between the Amazon EVS control plane and VPC. The default security group is used if a custom security group isn't specified.

The security group should allow access to the following.

TCP/UDP access to the DNS servers

HTTPS/SSH access to the host management VLAN subnet

HTTPS/SSH access to the Management VM VLAN subnet

You should avoid modifying the security group rules after deployment, as this can break the persistent connection between the Amazon EVS control plane and VPC. This can cause future environment actions like adding or removing hosts to fail.

"tags": Metadata that assists with categorization and organization. Each tag consists of a key and an optional value. You define both. Tags don't propagate to any other cluster or Amazon Web Services resources.

Main.Evs.create_environment_host — Function

create_environment_host(environment_id, host)
create_environment_host(environment_id, host, params::Dict{String,<:Any})

Creates an ESX host and adds it to an Amazon EVS environment. Amazon EVS supports 4-16 hosts per environment.

This action can only be used after the Amazon EVS environment is deployed.

You can use the dedicatedHostId parameter to specify an Amazon EC2 Dedicated Host for ESX host creation.

You can use the placementGroupId parameter to specify a cluster or partition placement group to launch EC2 instances into.

Note

If you don't specify an ESX version when adding hosts using CreateEnvironmentHost action, Amazon EVS automatically uses the default ESX version associated with your environment's VCF version. To find the default ESX version for a particular VCF version, use the GetVersions action.

Note

You cannot use the dedicatedHostId and placementGroupId parameters together in the same CreateEnvironmentHost action. This results in a ValidationException response.

Arguments

environment_id: A unique ID for the environment that the host is added to.

host: The host that is created and added to the environment.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": !!! note This parameter is not used in Amazon EVS currently. If you supply input for this parameter, it will have no effect.

A unique, case-sensitive identifier that you provide to ensure the idempotency of the host creation request. If you do not specify a client token, a randomly generated token is used for the request to ensure idempotency.

"esxVersion": The ESX version to use for the host.

Main.Evs.delete_environment — Function

delete_environment(environment_id)
delete_environment(environment_id, params::Dict{String,<:Any})

Deletes an Amazon EVS environment.

Amazon EVS environments will only be enabled for deletion once the hosts are deleted. You can delete hosts using the DeleteEnvironmentHost action.

Environment deletion also deletes the associated Amazon EVS VLAN subnets and Amazon Web Services Secrets Manager secrets that Amazon EVS created. Amazon Web Services resources that you create are not deleted. These resources may continue to incur costs.

Arguments

environment_id: A unique ID associated with the environment to be deleted.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": !!! note This parameter is not used in Amazon EVS currently. If you supply input for this parameter, it will have no effect.

A unique, case-sensitive identifier that you provide to ensure the idempotency of the environment deletion request. If you do not specify a client token, a randomly generated token is used for the request to ensure idempotency.

Main.Evs.delete_environment_host — Function

delete_environment_host(environment_id, host_name)
delete_environment_host(environment_id, host_name, params::Dict{String,<:Any})

Deletes a host from an Amazon EVS environment.

Note

Before deleting a host, you must unassign and decommission the host from within the SDDC Manager user interface. Not doing so could impact the availability of your virtual machines or result in data loss.

Arguments

environment_id: A unique ID for the host's environment.

host_name: The DNS hostname associated with the host to be deleted.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": !!! note This parameter is not used in Amazon EVS currently. If you supply input for this parameter, it will have no effect.

A unique, case-sensitive identifier that you provide to ensure the idempotency of the host deletion request. If you do not specify a client token, a randomly generated token is used for the request to ensure idempotency.

Main.Evs.disassociate_eip_from_vlan — Function

disassociate_eip_from_vlan(association_id, environment_id, vlan_name)
disassociate_eip_from_vlan(association_id, environment_id, vlan_name, params::Dict{String,<:Any})

Disassociates an Elastic IP address from a public HCX VLAN. This operation is only allowed for public HCX VLANs at this time.

Arguments

association_id: A unique ID for the Elastic IP address association.

environment_id: A unique ID for the environment containing the VLAN that the Elastic IP address disassociates from.

vlan_name: The name of the VLAN. hcx is the only accepted VLAN name at this time.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": !!! note This parameter is not used in Amazon EVS currently. If you supply input for this parameter, it will have no effect.

A unique, case-sensitive identifier that you provide to ensure the idempotency of the environment creation request. If you do not specify a client token, a randomly generated token is used for the request to ensure idempotency.

Main.Evs.get_environment — Function

get_environment(environment_id)
get_environment(environment_id, params::Dict{String,<:Any})

Returns a description of the specified environment.

Arguments

environment_id: A unique ID for the environment.

Main.Evs.get_versions — Function

get_versions()
get_versions(params::Dict{String,<:Any})

Returns information about VCF versions, ESX versions and EC2 instance types provided by Amazon EVS. For each VCF version, the response also includes the default ESX version and provided EC2 instance types.

Main.Evs.list_environment_hosts — Function

list_environment_hosts(environment_id)
list_environment_hosts(environment_id, params::Dict{String,<:Any})

List the hosts within an environment.

Arguments

environment_id: A unique ID for the environment.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return. If you specify MaxResults in the request, the response includes information up to the limit specified.

"nextToken": A unique pagination token for each page. If nextToken is returned, there are more results available. Make the call again using the returned token with all other arguments unchanged to retrieve the next page. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

Main.Evs.list_environment_vlans — Function

list_environment_vlans(environment_id)
list_environment_vlans(environment_id, params::Dict{String,<:Any})

Lists environment VLANs that are associated with the specified environment.

Arguments

environment_id: A unique ID for the environment.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return. If you specify MaxResults in the request, the response includes information up to the limit specified.

"nextToken": A unique pagination token for each page. If nextToken is returned, there are more results available. Make the call again using the returned token with all other arguments unchanged to retrieve the next page. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

Main.Evs.list_environments — Function

list_environments()
list_environments(params::Dict{String,<:Any})

Lists the Amazon EVS environments in your Amazon Web Services account in the specified Amazon Web Services Region.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return. If you specify MaxResults in the request, the response includes information up to the limit specified.

"nextToken": A unique pagination token for each page. If nextToken is returned, there are more results available. Make the call again using the returned token with all other arguments unchanged to retrieve the next page. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

"state": The state of an environment. Used to filter response results to return only environments with the specified environmentState.

Main.Evs.list_tags_for_resource — Function

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists the tags for an Amazon EVS resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) that identifies the resource to list tags for.

Main.Evs.tag_resource — Function

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Associates the specified tags to an Amazon EVS resource with the specified resourceArn. If existing tags on a resource are not specified in the request parameters, they aren't changed. When a resource is deleted, the tags associated with that resource are also deleted. Tags that you create for Amazon EVS resources don't propagate to any other resources associated with the environment. For example, if you tag an environment with this operation, that tag doesn't automatically propagate to the VLAN subnets and hosts associated with the environment.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource to add tags to.

tags: Metadata that assists with categorization and organization. Each tag consists of a key and an optional value. You define both. Tags don't propagate to any other environment or Amazon Web Services resources.

Main.Evs.untag_resource — Function

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Deletes specified tags from an Amazon EVS resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource to delete tags from.

tag_keys: The keys of the tags to delete.