Gameliftstreams

This page documents function available when using the Gameliftstreams module, created with @service Gameliftstreams.

Index

Main.Gameliftstreams.add_stream_group_locations

Main.Gameliftstreams.associate_applications

Main.Gameliftstreams.create_application

Main.Gameliftstreams.create_stream_group

Main.Gameliftstreams.create_stream_session_connection

Main.Gameliftstreams.delete_application

Main.Gameliftstreams.delete_stream_group

Main.Gameliftstreams.disassociate_applications

Main.Gameliftstreams.export_stream_session_files

Main.Gameliftstreams.get_application

Main.Gameliftstreams.get_stream_group

Main.Gameliftstreams.get_stream_session

Main.Gameliftstreams.list_applications

Main.Gameliftstreams.list_stream_groups

Main.Gameliftstreams.list_stream_sessions

Main.Gameliftstreams.list_stream_sessions_by_account

Main.Gameliftstreams.list_tags_for_resource

Main.Gameliftstreams.remove_stream_group_locations

Main.Gameliftstreams.start_stream_session

Main.Gameliftstreams.tag_resource

Main.Gameliftstreams.terminate_stream_session

Main.Gameliftstreams.untag_resource

Main.Gameliftstreams.update_application

Main.Gameliftstreams.update_stream_group

Documentation

Main.Gameliftstreams.add_stream_group_locations — Function

add_stream_group_locations(identifier, location_configurations)
add_stream_group_locations(identifier, location_configurations, params::Dict{String,<:Any})

Add locations that can host stream sessions. To add a location, the stream group must be in ACTIVE status. You configure locations and their corresponding capacity for each stream group. Creating a stream group in a location that's nearest to your end users can help minimize latency and improve quality.

This operation provisions stream capacity at the specified locations. By default, all locations have 1 or 2 capacity, depending on the stream class option: 2 for 'High' and 1 for 'Ultra' and 'Win2022'. This operation also copies the content files of all associated applications to an internal S3 bucket at each location. This allows Amazon GameLift Streams to host performant stream sessions.

Arguments

identifier: A stream group to add the specified locations to.

This value is an Amazon Resource Name (ARN) or ID that uniquely identifies the stream group resource. Example ARN: arn:aws:gameliftstreams:us-west-2:111122223333:streamgroup/sg-1AB2C3De4. Example ID: sg-1AB2C3De4.

location_configurations: A set of one or more locations and the streaming capacity for each location.

Main.Gameliftstreams.associate_applications — Function

associate_applications(application_identifiers, identifier)
associate_applications(application_identifiers, identifier, params::Dict{String,<:Any})

When you associate, or link, an application with a stream group, then Amazon GameLift Streams can launch the application using the stream group's allocated compute resources. The stream group must be in ACTIVE status. You can reverse this action by using DisassociateApplications.

If a stream group does not already have a linked application, Amazon GameLift Streams will automatically assign the first application provided in ApplicationIdentifiers as the default.

Arguments

application_identifiers: A set of applications to associate with the stream group.

This value is a set of either Amazon Resource Names (ARN) or IDs that uniquely identify application resources. Example ARN: arn:aws:gameliftstreams:us-west-2:111122223333:application/a-9ZY8X7Wv6. Example ID: a-9ZY8X7Wv6.

identifier: A stream group to associate to the applications.

This value is an Amazon Resource Name (ARN) or ID that uniquely identifies the stream group resource. Example ARN: arn:aws:gameliftstreams:us-west-2:111122223333:streamgroup/sg-1AB2C3De4. Example ID: sg-1AB2C3De4.

Main.Gameliftstreams.create_application — Function

create_application(application_source_uri, description, executable_path, runtime_environment)
create_application(application_source_uri, description, executable_path, runtime_environment, params::Dict{String,<:Any})

Creates an application resource in Amazon GameLift Streams, which specifies the application content you want to stream, such as a game build or other software, and configures the settings to run it.

Before you create an application, upload your application content files to an Amazon Simple Storage Service (Amazon S3) bucket. For more information, see Getting Started in the Amazon GameLift Streams Developer Guide.

Important

Make sure that your files in the Amazon S3 bucket are the correct version you want to use. If you change the files at a later time, you will need to create a new Amazon GameLift Streams application.

If the request is successful, Amazon GameLift Streams begins to create an application and sets the status to INITIALIZED. When an application reaches READY status, you can use the application to set up stream groups and start streams. To track application status, call GetApplication.

Arguments

application_source_uri: The location of the content that you want to stream. Enter an Amazon S3 URI to a bucket that contains your game or other application. The location can have a multi-level prefix structure, but it must include all the files needed to run the content. Amazon GameLift Streams copies everything under the specified location.

This value is immutable. To designate a different content location, create a new application.

Note

The Amazon S3 bucket and the Amazon GameLift Streams application must be in the same Amazon Web Services Region.

description: A human-readable label for the application. You can update this value later.

executable_path: The relative path and file name of the executable file that Amazon GameLift Streams will stream. Specify a path relative to the location set in ApplicationSourceUri. The file must be contained within the application's root folder. For Windows applications, the file must be a valid Windows executable or batch file with a filename ending in .exe, .cmd, or .bat. For Linux applications, the file must be a valid Linux binary executable or a script that contains an initial interpreter line starting with a shebang ('#!').

runtime_environment: Configuration settings that identify the operating system for an application resource. This can also include a compatibility layer and other drivers.

A runtime environment can be one of the following:

For Linux applications

Ubuntu 22.04 LTS (Type=UBUNTU, Version=22_04_LTS)

For Windows applications

Microsoft Windows Server 2022 Base (Type=WINDOWS, Version=2022)

Proton 9.0-2 (Type=PROTON, Version=20250516)

Proton 8.0-5 (Type=PROTON, Version=20241007)

Proton 8.0-2c (Type=PROTON, Version=20230704)

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ApplicationLogOutputUri": An Amazon S3 URI to a bucket where you would like Amazon GameLift Streams to save application logs. Required if you specify one or more ApplicationLogPaths.

Note

The log bucket must have permissions that give Amazon GameLift Streams access to write the log files. For more information, see Application log bucket permission policy in the Amazon GameLift Streams Developer Guide.

"ApplicationLogPaths": Locations of log files that your content generates during a stream session. Enter path values that are relative to the ApplicationSourceUri location. You can specify up to 10 log paths. Amazon GameLift Streams uploads designated log files to the Amazon S3 bucket that you specify in ApplicationLogOutputUri at the end of a stream session. To retrieve stored log files, call GetStreamSession and get the LogFileLocationUri.

"ClientToken": A unique identifier that represents a client request. The request is idempotent, which ensures that an API request completes only once. When users send a request, Amazon GameLift Streams automatically populates this field.

"Tags": A list of labels to assign to the new application resource. Tags are developer- defined key-value pairs. Tagging Amazon Web Services resources is useful for resource management, access management and cost allocation. See Tagging Amazon Web Services Resources in the Amazon Web Services General Reference. You can use TagResource to add tags, UntagResource to remove tags, and ListTagsForResource to view tags on existing resources.

Main.Gameliftstreams.create_stream_group — Function

create_stream_group(description, stream_class)
create_stream_group(description, stream_class, params::Dict{String,<:Any})

Stream groups manage how Amazon GameLift Streams allocates resources and handles concurrent streams, allowing you to effectively manage capacity and costs. Within a stream group, you specify an application to stream, streaming locations and their capacity, and the stream class you want to use when streaming applications to your end-users. A stream class defines the hardware configuration of the compute resources that Amazon GameLift Streams will use when streaming, such as the CPU, GPU, and memory.

Stream capacity represents the number of concurrent streams that can be active at a time. You set stream capacity per location, per stream group. The following capacity settings are available:

Always-on capacity: This setting, if non-zero, indicates minimum streaming capacity which is allocated to you and is never released back to the service. You pay for this base level of capacity at all times, whether used or idle.

Maximum capacity: This indicates the maximum capacity that the service can allocate for you. Newly created streams may take a few minutes to start. Capacity is released back to the service when idle. You pay for capacity that is allocated to you until it is released.

Target-idle capacity: This indicates idle capacity which the service pre-allocates and holds for you in anticipation of future activity. This helps to insulate your users from capacity-allocation delays. You pay for capacity which is held in this intentional idle state.

Values for capacity must be whole number multiples of the tenancy value of the stream group's stream class.

To adjust the capacity of any ACTIVE stream group, call UpdateStreamGroup.

If the CreateStreamGroup request is successful, Amazon GameLift Streams assigns a unique ID to the stream group resource and sets the status to ACTIVATING. It can take a few minutes for Amazon GameLift Streams to finish creating the stream group while it searches for unallocated compute resources and provisions them. When complete, the stream group status will be ACTIVE and you can start stream sessions by using StartStreamSession. To check the stream group's status, call GetStreamGroup.

Stream groups should be recreated every 3-4 weeks to pick up important service updates and fixes. Stream groups that are older than 180 days can no longer be updated with new application associations. Stream groups expire when they are 365 days old, at which point they can no longer stream sessions. The exact expiration date is indicated by the date value in the ExpiresAt field.

Arguments

description: A descriptive label for the stream group.

stream_class: The target stream quality for sessions that are hosted in this stream group. Set a stream class that is appropriate to the type of content that you're streaming. Stream class determines the type of computing resources Amazon GameLift Streams uses and impacts the cost of streaming. The following options are available:

A stream class can be one of the following:

gen6n_pro_win2022 (NVIDIA, pro) Supports applications with extremely high 3D scene complexity which require maximum resources. Runs applications on Microsoft Windows Server 2022 Base and supports DirectX 12. Compatible with Unreal Engine versions up through 5.6, 32 and 64-bit applications, and anti-cheat technology. Uses NVIDIA L4 Tensor Core GPU.

Reference resolution: 1080p

Reference frame rate: 60 fps

Workload specifications: 16 vCPUs, 64 GB RAM, 24 GB VRAM

Tenancy: Supports 1 concurrent stream session

gen6n_pro (NVIDIA, pro) Supports applications with extremely high 3D scene complexity which require maximum resources. Uses dedicated NVIDIA L4 Tensor Core GPU.

Reference resolution: 1080p

Reference frame rate: 60 fps

Workload specifications: 16 vCPUs, 64 GB RAM, 24 GB VRAM

Tenancy: Supports 1 concurrent stream session

gen6n_ultra_win2022 (NVIDIA, ultra) Supports applications with high 3D scene complexity. Runs applications on Microsoft Windows Server 2022 Base and supports DirectX

Compatible with Unreal Engine versions up through 5.6, 32 and 64-bit applications,

and anti-cheat technology. Uses NVIDIA L4 Tensor Core GPU. - Reference resolution: 1080p - Reference frame rate: 60 fps - Workload specifications: 8 vCPUs, 32 GB RAM, 24 GB VRAM - Tenancy: Supports 1 concurrent stream session

gen6n_ultra (NVIDIA, ultra) Supports applications with high 3D scene complexity. Uses dedicated NVIDIA L4 Tensor Core GPU.

Reference resolution: 1080p

Reference frame rate: 60 fps

Workload specifications: 8 vCPUs, 32 GB RAM, 24 GB VRAM

Tenancy: Supports 1 concurrent stream session

gen6n_high (NVIDIA, high) Supports applications with moderate to high 3D scene complexity. Uses NVIDIA L4 Tensor Core GPU.

Reference resolution: 1080p

Reference frame rate: 60 fps

Workload specifications: 4 vCPUs, 16 GB RAM, 12 GB VRAM

Tenancy: Supports up to 2 concurrent stream sessions

gen6n_medium (NVIDIA, medium) Supports applications with moderate 3D scene complexity. Uses NVIDIA L4 Tensor Core GPU.

Reference resolution: 1080p

Reference frame rate: 60 fps

Workload specifications: 2 vCPUs, 8 GB RAM, 6 GB VRAM

Tenancy: Supports up to 4 concurrent stream sessions

gen6n_small (NVIDIA, small) Supports applications with lightweight 3D scene complexity and low CPU usage. Uses NVIDIA L4 Tensor Core GPU.

Reference resolution: 1080p

Reference frame rate: 60 fps

Workload specifications: 1 vCPUs, 4 GB RAM, 2 GB VRAM

Tenancy: Supports up to 12 concurrent stream sessions

gen5n_win2022 (NVIDIA, ultra) Supports applications with extremely high 3D scene complexity. Runs applications on Microsoft Windows Server 2022 Base and supports DirectX

Compatible with Unreal Engine versions up through 5.6, 32 and 64-bit applications,

and anti-cheat technology. Uses NVIDIA A10G Tensor Core GPU. - Reference resolution: 1080p - Reference frame rate: 60 fps - Workload specifications: 8 vCPUs, 32 GB RAM, 24 GB VRAM - Tenancy: Supports 1 concurrent stream session

gen5n_high (NVIDIA, high) Supports applications with moderate to high 3D scene complexity. Uses NVIDIA A10G Tensor Core GPU.

Reference resolution: 1080p

Reference frame rate: 60 fps

Workload specifications: 4 vCPUs, 16 GB RAM, 12 GB VRAM

Tenancy: Supports up to 2 concurrent stream sessions

gen5n_ultra (NVIDIA, ultra) Supports applications with extremely high 3D scene complexity. Uses dedicated NVIDIA A10G Tensor Core GPU.

Reference resolution: 1080p

Reference frame rate: 60 fps

Workload specifications: 8 vCPUs, 32 GB RAM, 24 GB VRAM

Tenancy: Supports 1 concurrent stream session

gen4n_win2022 (NVIDIA, ultra) Supports applications with extremely high 3D scene complexity. Runs applications on Microsoft Windows Server 2022 Base and supports DirectX

Compatible with Unreal Engine versions up through 5.6, 32 and 64-bit applications,

and anti-cheat technology. Uses NVIDIA T4 Tensor Core GPU. - Reference resolution: 1080p - Reference frame rate: 60 fps - Workload specifications: 8 vCPUs, 32 GB RAM, 16 GB VRAM - Tenancy: Supports 1 concurrent stream session

gen4n_high (NVIDIA, high) Supports applications with moderate to high 3D scene complexity. Uses NVIDIA T4 Tensor Core GPU.

Reference resolution: 1080p

Reference frame rate: 60 fps

Workload specifications: 4 vCPUs, 16 GB RAM, 8 GB VRAM

Tenancy: Supports up to 2 concurrent stream sessions

gen4n_ultra (NVIDIA, ultra) Supports applications with high 3D scene complexity. Uses dedicated NVIDIA T4 Tensor Core GPU.

Reference resolution: 1080p

Reference frame rate: 60 fps

Workload specifications: 8 vCPUs, 32 GB RAM, 16 GB VRAM

Tenancy: Supports 1 concurrent stream session

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": A unique identifier that represents a client request. The request is idempotent, which ensures that an API request completes only once. When users send a request, Amazon GameLift Streams automatically populates this field.

"DefaultApplicationIdentifier": The unique identifier of the Amazon GameLift Streams application that you want to set as the default application in a stream group. The application that you specify must be in READY status. The default application is pre- cached on always-on compute resources, reducing stream startup times. Other applications are automatically cached as needed.

If you do not link an application when you create a stream group, you will need to link one later, before you can start streaming, using AssociateApplications.

This value is an Amazon Resource Name (ARN) or ID that uniquely identifies the application resource. Example ARN: arn:aws:gameliftstreams:us-west-2:111122223333:application/a-9ZY8X7Wv6. Example ID: a-9ZY8X7Wv6.

"LocationConfigurations": A set of one or more locations and the streaming capacity for each location.

"Tags": A list of labels to assign to the new stream group resource. Tags are developer- defined key-value pairs. Tagging Amazon Web Services resources is useful for resource management, access management and cost allocation. See Tagging Amazon Web Services Resources in the Amazon Web Services General Reference. You can use TagResource to add tags, UntagResource to remove tags, and ListTagsForResource to view tags on existing resources.

Main.Gameliftstreams.create_stream_session_connection — Function

create_stream_session_connection(identifier, signal_request, stream_session_identifier)
create_stream_session_connection(identifier, signal_request, stream_session_identifier, params::Dict{String,<:Any})

Enables clients to reconnect to a stream session while preserving all session state and data in the disconnected session. This reconnection process can be initiated when a stream session is in either PENDING_CLIENT_RECONNECTION or ACTIVE status. The process works as follows:

Initial disconnect:

When a client disconnects or loses connection, the stream session transitions from CONNECTED to PENDING_CLIENT_RECONNECTION

Reconnection time window:

Clients have ConnectionTimeoutSeconds (defined in StartStreamSession) to reconnect before session termination

Your backend server must call CreateStreamSessionConnection to initiate reconnection

Session transitions to RECONNECTING status

Reconnection completion:

On successful CreateStreamSessionConnection, session status changes to ACTIVE

Provide the new connection information to the requesting client

Client must establish connection within ConnectionTimeoutSeconds

Session terminates automatically if client fails to connect in time

For more information about the stream session lifecycle, see Stream sessions in the Amazon GameLift Streams Developer Guide.

To begin re-connecting to an existing stream session, specify the stream group ID and stream session ID that you want to reconnect to, and the signal request to use with the stream.

Arguments

identifier: Amazon Resource Name (ARN) or ID that uniquely identifies the stream group resource. Example ARN: arn:aws:gameliftstreams:us-west-2:111122223333:streamgroup/sg-1AB2C3De4. Example ID: sg-1AB2C3De4.

The stream group that you want to run this stream session with. The stream group must be in ACTIVE status.

signal_request: A WebRTC ICE offer string to use when initializing a WebRTC connection. The offer is a very long JSON string. Provide the string as a text value in quotes. The offer must be newly generated, not the same offer provided to StartStreamSession.

stream_session_identifier: Amazon Resource Name (ARN) or ID that uniquely identifies the stream session resource. Example ARN: arn:aws:gameliftstreams:us-west-2:111122223333:streamsession/sg-1AB2C3De4/ABC123def4567. Example ID: ABC123def4567.

The stream session must be in PENDING_CLIENT_RECONNECTION or ACTIVE status.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": A unique identifier that represents a client request. The request is idempotent, which ensures that an API request completes only once. When users send a request, Amazon GameLift Streams automatically populates this field.

Main.Gameliftstreams.delete_application — Function

delete_application(identifier)
delete_application(identifier, params::Dict{String,<:Any})

Permanently deletes an Amazon GameLift Streams application resource. This also deletes the application content files stored with Amazon GameLift Streams. However, this does not delete the original files that you uploaded to your Amazon S3 bucket; you can delete these any time after Amazon GameLift Streams creates an application, which is the only time Amazon GameLift Streams accesses your Amazon S3 bucket.

You can only delete an application that meets the following conditions:

The application is in READY or ERROR status. You cannot delete an application that's in PROCESSING or INITIALIZED status.

The application is not the default application of any stream groups. You must first delete the stream group by using DeleteStreamGroup.

The application is not linked to any stream groups. You must first unlink the stream group by using DisassociateApplications.

An application is not streaming in any ongoing stream session. You must wait until the client ends the stream session or call TerminateStreamSession to end the stream.

If any active stream groups exist for this application, this request returns a ValidationException.

Arguments

identifier: An Amazon Resource Name (ARN) or ID that uniquely identifies the application resource. Example ARN: arn:aws:gameliftstreams:us-west-2:111122223333:application/a-9ZY8X7Wv6. Example ID: a-9ZY8X7Wv6.

Main.Gameliftstreams.delete_stream_group — Function

delete_stream_group(identifier)
delete_stream_group(identifier, params::Dict{String,<:Any})

Permanently deletes all compute resources and information related to a stream group. To delete a stream group, specify the unique stream group identifier. During the deletion process, the stream group's status is DELETING. This operation stops streams in progress and prevents new streams from starting. As a best practice, before deleting the stream group, call ListStreamSessions to check for streams in progress and take action to stop them. When you delete a stream group, any application associations referring to that stream group are automatically removed.

Arguments

identifier: An Amazon Resource Name (ARN) or ID that uniquely identifies the stream group resource. Example ARN: arn:aws:gameliftstreams:us-west-2:111122223333:streamgroup/sg-1AB2C3De4. Example ID: sg-1AB2C3De4.

Main.Gameliftstreams.disassociate_applications — Function

disassociate_applications(application_identifiers, identifier)
disassociate_applications(application_identifiers, identifier, params::Dict{String,<:Any})

When you disassociate, or unlink, an application from a stream group, you can no longer stream this application by using that stream group's allocated compute resources. Any streams in process will continue until they terminate, which helps avoid interrupting an end-user's stream. Amazon GameLift Streams will not initiate new streams in the stream group using the disassociated application. The disassociate action does not affect the stream capacity of a stream group. To disassociate an application, the stream group must be in ACTIVE status.

If you disassociate the default application, Amazon GameLift Streams will automatically choose a new default application from the remaining associated applications. To change which application is the default application, call UpdateStreamGroup and specify a new DefaultApplicationIdentifier.

Arguments

application_identifiers: A set of applications that you want to disassociate from the stream group.

This value is a set of either Amazon Resource Names (ARN) or IDs that uniquely identify application resources. Example ARN: arn:aws:gameliftstreams:us-west-2:111122223333:application/a-9ZY8X7Wv6. Example ID: a-9ZY8X7Wv6.

identifier: A stream group to disassociate these applications from.

This value is an Amazon Resource Name (ARN) or ID that uniquely identifies the stream group resource. Example ARN: arn:aws:gameliftstreams:us-west-2:111122223333:streamgroup/sg-1AB2C3De4. Example ID: sg-1AB2C3De4.

Main.Gameliftstreams.export_stream_session_files — Function

export_stream_session_files(identifier, output_uri, stream_session_identifier)
export_stream_session_files(identifier, output_uri, stream_session_identifier, params::Dict{String,<:Any})

Export the files that your application modifies or generates in a stream session, which can help you debug or verify your application. When your application runs, it generates output files such as logs, diagnostic information, crash dumps, save files, user data, screenshots, and so on. The files can be defined by the engine or frameworks that your application uses, or information that you've programmed your application to output.

You can only call this action on a stream session that is in progress, specifically in one of the following statuses ACTIVE, CONNECTED, PENDING_CLIENT_RECONNECTION, and RECONNECTING. You must provide an Amazon Simple Storage Service (Amazon S3) bucket to store the files in. When the session ends, Amazon GameLift Streams produces a compressed folder that contains all of the files and directories that were modified or created by the application during the stream session. AWS uses your security credentials to authenticate and authorize access to your Amazon S3 bucket.

Amazon GameLift Streams collects the following generated and modified files. Find them in the corresponding folders in the .zip archive.

application/: The folder where your application or game is stored.

profile/: The user profile folder.

temp/: The system temp folder.

To verify the status of the exported files, use GetStreamSession.

To delete the files, delete the object in the S3 bucket.

Arguments

identifier: An Amazon Resource Name (ARN) or ID that uniquely identifies the stream group resource. Example ARN: arn:aws:gameliftstreams:us-west-2:111122223333:streamgroup/sg-1AB2C3De4. Example ID: sg-1AB2C3De4.

output_uri: The S3 bucket URI where Amazon GameLift Streams uploads the set of compressed exported files for this stream session. Amazon GameLift Streams generates a ZIP file name based on the stream session metadata. Alternatively, you can provide a custom file name with a .zip file extension.

Example 1: If you provide an S3 URI called s3://amzn-s3-demo-destination-bucket/MyGame_Session1.zip, then Amazon GameLift Streams will save the files at that location.

Example 2: If you provide an S3 URI called s3://amzn-s3-demo-destination-bucket/MyGameSessions_ExportedFiles/, then Amazon GameLift Streams will save the files at s3://amzn-s3-demo-destination-bucket/MyGameSessions_ExportedFiles/YYYYMMDD-HHMMSS-appId-sg-Id-sessionId.zip or another similar name.

stream_session_identifier: An Amazon Resource Name (ARN) or ID that uniquely identifies the stream session resource. Example ARN: arn:aws:gameliftstreams:us-west-2:111122223333:streamsession/sg-1AB2C3De4/ABC123def4567. Example ID: ABC123def4567.

Main.Gameliftstreams.get_application — Function

get_application(identifier)
get_application(identifier, params::Dict{String,<:Any})

Retrieves properties for an Amazon GameLift Streams application resource. Specify the ID of the application that you want to retrieve. If the operation is successful, it returns properties for the requested application.

Arguments

identifier: An Amazon Resource Name (ARN) or ID that uniquely identifies the application resource. Example ARN: arn:aws:gameliftstreams:us-west-2:111122223333:application/a-9ZY8X7Wv6. Example ID: a-9ZY8X7Wv6.

Main.Gameliftstreams.get_stream_group — Function

get_stream_group(identifier)
get_stream_group(identifier, params::Dict{String,<:Any})

Retrieves properties for a Amazon GameLift Streams stream group resource. Specify the ID of the stream group that you want to retrieve. If the operation is successful, it returns properties for the requested stream group.

Arguments

identifier: An Amazon Resource Name (ARN) or ID that uniquely identifies the stream group resource. Example ARN: arn:aws:gameliftstreams:us-west-2:111122223333:streamgroup/sg-1AB2C3De4. Example ID: sg-1AB2C3De4.

Main.Gameliftstreams.get_stream_session — Function

get_stream_session(identifier, stream_session_identifier)
get_stream_session(identifier, stream_session_identifier, params::Dict{String,<:Any})

Retrieves properties for a Amazon GameLift Streams stream session resource. Specify the Amazon Resource Name (ARN) of the stream session that you want to retrieve and its stream group ARN. If the operation is successful, it returns properties for the requested resource.

Arguments

identifier: The stream group that runs this stream session.

This value is an Amazon Resource Name (ARN) or ID that uniquely identifies the stream group resource. Example ARN: arn:aws:gameliftstreams:us-west-2:111122223333:streamgroup/sg-1AB2C3De4. Example ID: sg-1AB2C3De4.

stream_session_identifier: An Amazon Resource Name (ARN) or ID that uniquely identifies the stream session resource. Example ARN: arn:aws:gameliftstreams:us-west-2:111122223333:streamsession/sg-1AB2C3De4/ABC123def4567. Example ID: ABC123def4567.

Main.Gameliftstreams.list_applications — Function

list_applications()
list_applications(params::Dict{String,<:Any})

Retrieves a list of all Amazon GameLift Streams applications that are associated with the Amazon Web Services account in use. This operation returns applications in all statuses, in no particular order. You can paginate the results as needed.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The number of results to return. Use this parameter with NextToken to return results in sequential pages. Default value is 25.

"NextToken": The token that marks the start of the next set of results. Use this token when you retrieve results as sequential pages. To get the first page of results, omit a token value. To get the remaining pages, provide the token returned with the previous result set.

Main.Gameliftstreams.list_stream_groups — Function

list_stream_groups()
list_stream_groups(params::Dict{String,<:Any})

Retrieves a list of all Amazon GameLift Streams stream groups that are associated with the Amazon Web Services account in use. This operation returns stream groups in all statuses, in no particular order. You can paginate the results as needed.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The number of results to return. Use this parameter with NextToken to return results in sequential pages. Default value is 25.

"NextToken": A token that marks the start of the next set of results. Use this token when you retrieve results as sequential pages. To get the first page of results, omit a token value. To get the remaining pages, provide the token returned with the previous result set.

Main.Gameliftstreams.list_stream_sessions — Function

list_stream_sessions(identifier)
list_stream_sessions(identifier, params::Dict{String,<:Any})

Retrieves a list of Amazon GameLift Streams stream sessions that a stream group is hosting.

To retrieve stream sessions, specify the stream group, and optionally filter by stream session status. You can paginate the results as needed.

This operation returns the requested stream sessions in no particular order.

Arguments

identifier: The unique identifier of a Amazon GameLift Streams stream group to retrieve the stream session for. You can use either the stream group ID or the Amazon Resource Name (ARN).

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ExportFilesStatus": Filter by the exported files status. You can specify one status in each request to retrieve only sessions that currently have that exported files status.

Exported files can be in one of the following states:

SUCCEEDED: The exported files are successfully stored in an S3 bucket.

FAILED: The session ended but Amazon GameLift Streams couldn't collect and upload the files to S3.

PENDING: Either the stream session is still in progress, or uploading the exported files to the S3 bucket is in progress.

"MaxResults": The number of results to return. Use this parameter with NextToken to return results in sequential pages. Default value is 25.

"NextToken": The token that marks the start of the next set of results. Use this token when you retrieve results as sequential pages. To get the first page of results, omit a token value. To get the remaining pages, provide the token returned with the previous result set.

"Status": Filter by the stream session status. You can specify one status in each request to retrieve only sessions that are currently in that status.

Main.Gameliftstreams.list_stream_sessions_by_account — Function

list_stream_sessions_by_account()
list_stream_sessions_by_account(params::Dict{String,<:Any})

Retrieves a list of Amazon GameLift Streams stream sessions that this user account has access to.

In the returned list of stream sessions, the ExportFilesMetadata property only shows the Status value. To get the OutpurUri and StatusReason values, use GetStreamSession.

We don't recommend using this operation to regularly check stream session statuses because it's costly. Instead, to check status updates for a specific stream session, use GetStreamSession.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ExportFilesStatus": Filter by the exported files status. You can specify one status in each request to retrieve only sessions that currently have that exported files status.

"MaxResults": The number of results to return. Use this parameter with NextToken to return results in sequential pages. Default value is 25.

"NextToken": The token that marks the start of the next set of results. Use this token when you retrieve results as sequential pages. To get the first page of results, omit a token value. To get the remaining pages, provide the token returned with the previous result set.

"Status": Filter by the stream session status. You can specify one status in each request to retrieve only sessions that are currently in that status.

Main.Gameliftstreams.list_tags_for_resource — Function

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Retrieves all tags assigned to a Amazon GameLift Streams resource. To list tags for a resource, specify the ARN value for the resource.

Learn more

Tagging Amazon Web Services Resources in the Amazon Web Services General Reference

Amazon Web Services Tagging Strategies

Arguments

resource_arn: The Amazon Resource Name (ARN) that you want to retrieve tags for. To get an Amazon GameLift Streams resource ARN, call a List or Get operation for the resource.

Main.Gameliftstreams.remove_stream_group_locations — Function

remove_stream_group_locations(identifier, locations)
remove_stream_group_locations(identifier, locations, params::Dict{String,<:Any})

Removes a set of remote locations from this stream group. To remove a location, the stream group must be in ACTIVE status. When you remove a location, Amazon GameLift Streams releases allocated compute resources in that location. Stream sessions can no longer start from removed locations in a stream group. Amazon GameLift Streams also deletes the content files of all associated applications that were in Amazon GameLift Streams's internal Amazon S3 bucket at this location.

You cannot remove the Amazon Web Services Region location where you initially created this stream group, known as the primary location. However, you can set the stream capacity to zero to avoid incurring costs for allocated compute resources in that location.

Arguments

identifier: A stream group to remove the specified locations from.

This value is an Amazon Resource Name (ARN) or ID that uniquely identifies the stream group resource. Example ARN: arn:aws:gameliftstreams:us-west-2:111122223333:streamgroup/sg-1AB2C3De4. Example ID: sg-1AB2C3De4.

locations: A set of locations to remove this stream group. For example, us-east-1.

For a complete list of locations that Amazon GameLift Streams supports, refer to Regions, quotas, and limitations in the Amazon GameLift Streams Developer Guide.

Main.Gameliftstreams.start_stream_session — Function

start_stream_session(application_identifier, identifier, protocol, signal_request)
start_stream_session(application_identifier, identifier, protocol, signal_request, params::Dict{String,<:Any})

This action initiates a new stream session and outputs connection information that clients can use to access the stream. A stream session refers to an instance of a stream that Amazon GameLift Streams transmits from the server to the end-user. A stream session runs on a compute resource that a stream group has allocated. The start stream session process works as follows:

Prerequisites:

You must have a stream group in ACTIVE status

You must have idle or on-demand capacity in a stream group in the location you want to stream from

You must have at least one application associated to the stream group (use AssociateApplications if needed)

Start stream request:

Your backend server calls StartStreamSession to initiate connection

Amazon GameLift Streams creates the stream session resource, assigns an Amazon Resource Name (ARN) value, and begins searching for available stream capacity to run the stream

Session transitions to ACTIVATING status

Placement completion:

If Amazon GameLift Streams is successful in finding capacity for the stream, the stream session status changes to ACTIVE status and StartStreamSession returns stream connection information

If Amazon GameLift Streams was not successful in finding capacity within the placement timeout period (defined according to the capacity type and platform type), the stream session status changes to ERROR status and StartStreamSession returns a StatusReason of placementTimeout

Connection completion:

Provide the new connection information to the requesting client

Client must establish connection within ConnectionTimeoutSeconds (specified in StartStreamSession parameters)

Session terminates automatically if client fails to connect in time

For more information about the stream session lifecycle, see Stream sessions in the Amazon GameLift Streams Developer Guide.

Timeouts to be aware of that affect a stream session:

Placement timeout: The amount of time that Amazon GameLift Streams has to find capacity for a stream request. Placement timeout varies based on the capacity type used to fulfill your stream request:

Always-on capacity: 75 seconds

On-demand capacity:

Linux/Proton runtimes: 90 seconds

Windows runtime: 10 minutes

Connection timeout: The amount of time that Amazon GameLift Streams waits for a client to connect to a stream session in ACTIVE status, or reconnect to a stream session in PENDING_CLIENT_RECONNECTION status, the latter of which occurs when a client disconnects or loses connection from a stream session. If no client connects before the timeout, Amazon GameLift Streams terminates the stream session. This value is specified by ConnectionTimeoutSeconds in the StartStreamSession parameters.

Idle timeout: A stream session will be terminated if no user input has been received for 60 minutes.

Maximum session length: A stream session will be terminated after this amount of time has elapsed since it started, regardless of any existing client connections. This value is specified by SessionLengthSeconds in the StartStreamSession parameters.

To start a new stream session, specify a stream group ID and application ID, along with the transport protocol and signal request to use with the stream session.

For stream groups that have multiple locations, provide a set of locations ordered by priority using a Locations parameter. Amazon GameLift Streams will start a single stream session in the next available location. An application must be finished replicating to a remote location before the remote location can host a stream.

To reconnect to a stream session after a client disconnects or loses connection, use CreateStreamSessionConnection.

Arguments

application_identifier: An Amazon Resource Name (ARN) or ID that uniquely identifies the application resource. Example ARN: arn:aws:gameliftstreams:us-west-2:111122223333:application/a-9ZY8X7Wv6. Example ID: a-9ZY8X7Wv6.

identifier: The stream group to run this stream session with.

This value is an Amazon Resource Name (ARN) or ID that uniquely identifies the stream group resource. Example ARN: arn:aws:gameliftstreams:us-west-2:111122223333:streamgroup/sg-1AB2C3De4. Example ID: sg-1AB2C3De4.

protocol: The data transport protocol to use for the stream session.

signal_request: A WebRTC ICE offer string to use when initializing a WebRTC connection. Typically, the offer is a very long JSON string. Provide the string as a text value in quotes.

Amazon GameLift Streams also supports setting the field to "NOCLIENTCONNECTION". This will create a session without needing any browser request or Web SDK integration. The session starts up as usual and waits for a reconnection from a browser, which is accomplished using CreateStreamSessionConnection.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AdditionalEnvironmentVariables": A set of options that you can use to control the stream session runtime environment, expressed as a set of key-value pairs. You can use this to configure the application or stream session details. You can also provide custom environment variables that Amazon GameLift Streams passes to your game client.

Note

If you want to debug your application with environment variables, we recommend that you do so in a local environment outside of Amazon GameLift Streams. For more information, refer to the Compatibility Guidance in the troubleshooting section of the Developer Guide.

AdditionalEnvironmentVariables and AdditionalLaunchArgs have similar purposes. AdditionalEnvironmentVariables passes data using environment variables; while AdditionalLaunchArgs passes data using command-line arguments.

"AdditionalLaunchArgs": A list of CLI arguments that are sent to the streaming server when a stream session launches. You can use this to configure the application or stream session details. You can also provide custom arguments that Amazon GameLift Streams passes to your game client.

AdditionalEnvironmentVariables and AdditionalLaunchArgs have similar purposes. AdditionalEnvironmentVariables passes data using environment variables; while AdditionalLaunchArgs passes data using command-line arguments.

"ClientToken": A unique identifier that represents a client request. The request is idempotent, which ensures that an API request completes only once. When users send a request, Amazon GameLift Streams automatically populates this field.

"ConnectionTimeoutSeconds": Length of time (in seconds) that Amazon GameLift Streams should wait for a client to connect or reconnect to the stream session. Applies to both connection and reconnection scenarios. This time span starts when the stream session reaches ACTIVE state. If no client connects before the timeout, Amazon GameLift Streams terminates the stream session. Default value is 120.

"Description": A human-readable label for the stream session. You can update this value later.

"Locations": A list of locations, in order of priority, where you want Amazon GameLift Streams to start a stream from. For example, us-east-1. Amazon GameLift Streams selects the location with the next available capacity to start a single stream session in. If this value is empty, Amazon GameLift Streams attempts to start a stream session in the primary location.

For a complete list of locations that Amazon GameLift Streams supports, refer to Regions, quotas, and limitations in the Amazon GameLift Streams Developer Guide.

"PerformanceStatsConfiguration": Configuration settings for sharing the stream session's performance stats with the client

"SessionLengthSeconds": The maximum duration of a session. Amazon GameLift Streams will automatically terminate a session after this amount of time has elapsed, regardless of any existing client connections. Default value is 43200 (12 hours).

"UserId": An opaque, unique identifier for an end-user, defined by the developer.

Main.Gameliftstreams.tag_resource — Function

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Assigns one or more tags to a Amazon GameLift Streams resource. Use tags to organize Amazon Web Services resources for a range of purposes. You can assign tags to the following Amazon GameLift Streams resource types:

Application

StreamGroup

Learn more

Tagging Amazon Web Services Resources in the Amazon Web Services General Reference

Amazon Web Services Tagging Strategies

Arguments

resource_arn: The Amazon Resource Name (ARN) of the Amazon GameLift Streams resource that you want to apply tags to.

tags: A list of tags, in the form of key-value pairs, to assign to the specified Amazon GameLift Streams resource.

Main.Gameliftstreams.terminate_stream_session — Function

terminate_stream_session(identifier, stream_session_identifier)
terminate_stream_session(identifier, stream_session_identifier, params::Dict{String,<:Any})

Permanently terminates an active stream session. When called, the stream session status changes to TERMINATING. You can terminate a stream session in any status except ACTIVATING. If the stream session is in ACTIVATING status, an exception is thrown.

Arguments

identifier: Amazon Resource Name (ARN) or ID that uniquely identifies the stream group resource. Example ARN: arn:aws:gameliftstreams:us-west-2:111122223333:streamgroup/sg-1AB2C3De4. Example ID: sg-1AB2C3De4.

The stream group that runs this stream session.

stream_session_identifier: Amazon Resource Name (ARN) or ID that uniquely identifies the stream session resource. Example ARN: arn:aws:gameliftstreams:us-west-2:111122223333:streamsession/sg-1AB2C3De4/ABC123def4567. Example ID: ABC123def4567.

Main.Gameliftstreams.untag_resource — Function

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes one or more tags from a Amazon GameLift Streams resource. To remove tags, specify the Amazon GameLift Streams resource and a list of one or more tags to remove.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the Amazon GameLift Streams resource that you want to remove tags from.

tag_keys: A list of tag keys to remove from the specified Amazon GameLift Streams resource.

Main.Gameliftstreams.update_application — Function

update_application(identifier)
update_application(identifier, params::Dict{String,<:Any})

Updates the mutable configuration settings for a Amazon GameLift Streams application resource. You can change the Description, ApplicationLogOutputUri, and ApplicationLogPaths.

To update application settings, specify the application ID and provide the new values. If the operation is successful, it returns the complete updated set of settings for the application.

Arguments

identifier: An Amazon Resource Name (ARN) or ID that uniquely identifies the application resource. Example ARN: arn:aws:gameliftstreams:us-west-2:111122223333:application/a-9ZY8X7Wv6. Example ID: a-9ZY8X7Wv6.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ApplicationLogOutputUri": An Amazon S3 URI to a bucket where you would like Amazon GameLift Streams to save application logs. Required if you specify one or more ApplicationLogPaths.

Note

The log bucket must have permissions that give Amazon GameLift Streams access to write the log files. For more information, see Application log bucket permission policy in the Amazon GameLift Streams Developer Guide.

"ApplicationLogPaths": Locations of log files that your content generates during a stream session. Enter path values that are relative to the ApplicationSourceUri location. You can specify up to 10 log paths. Amazon GameLift Streams uploads designated log files to the Amazon S3 bucket that you specify in ApplicationLogOutputUri at the end of a stream session. To retrieve stored log files, call GetStreamSession and get the LogFileLocationUri.

"Description": A human-readable label for the application.

Main.Gameliftstreams.update_stream_group — Function

update_stream_group(identifier)
update_stream_group(identifier, params::Dict{String,<:Any})

Updates the configuration settings for an Amazon GameLift Streams stream group resource. To update a stream group, it must be in ACTIVE status. You can change the description, the set of locations, and the requested capacity of a stream group per location. If you want to change the stream class, create a new stream group.

Stream capacity represents the number of concurrent streams that can be active at a time. You set stream capacity per location, per stream group. The following capacity settings are available:

Always-on capacity: This setting, if non-zero, indicates minimum streaming capacity which is allocated to you and is never released back to the service. You pay for this base level of capacity at all times, whether used or idle.

Maximum capacity: This indicates the maximum capacity that the service can allocate for you. Newly created streams may take a few minutes to start. Capacity is released back to the service when idle. You pay for capacity that is allocated to you until it is released.

Target-idle capacity: This indicates idle capacity which the service pre-allocates and holds for you in anticipation of future activity. This helps to insulate your users from capacity-allocation delays. You pay for capacity which is held in this intentional idle state.

Values for capacity must be whole number multiples of the tenancy value of the stream group's stream class.

To update a stream group, specify the stream group's Amazon Resource Name (ARN) and provide the new values. If the request is successful, Amazon GameLift Streams returns the complete updated metadata for the stream group. Expired stream groups cannot be updated.

Arguments

identifier: An Amazon Resource Name (ARN) or ID that uniquely identifies the stream group resource. Example ARN: arn:aws:gameliftstreams:us-west-2:111122223333:streamgroup/sg-1AB2C3De4. Example ID: sg-1AB2C3De4.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DefaultApplicationIdentifier": The unique identifier of the Amazon GameLift Streams application that you want to set as the default application in a stream group. The application that you specify must be in READY status. The default application is pre- cached on always-on compute resources, reducing stream startup times. Other applications are automatically cached as needed.

Note that this parameter only sets the default application in a stream group. To associate a new application to an existing stream group, you must use AssociateApplications.

When you switch default applications in a stream group, it can take up to a few hours for the new default application to be pre-cached.

This value is an Amazon Resource Name (ARN) or ID that uniquely identifies the application resource. Example ARN: arn:aws:gameliftstreams:us-west-2:111122223333:application/a-9ZY8X7Wv6. Example ID: a-9ZY8X7Wv6.

"Description": A descriptive label for the stream group.

"LocationConfigurations": A set of one or more locations and the streaming capacity for each location.