Gamesparks

This page documents function available when using the Gamesparks module, created with @service Gamesparks.

Index

Main.Gamesparks.create_game

Main.Gamesparks.create_snapshot

Main.Gamesparks.create_stage

Main.Gamesparks.delete_game

Main.Gamesparks.delete_stage

Main.Gamesparks.disconnect_player

Main.Gamesparks.export_snapshot

Main.Gamesparks.get_extension

Main.Gamesparks.get_extension_version

Main.Gamesparks.get_game

Main.Gamesparks.get_game_configuration

Main.Gamesparks.get_generated_code_job

Main.Gamesparks.get_player_connection_status

Main.Gamesparks.get_snapshot

Main.Gamesparks.get_stage

Main.Gamesparks.get_stage_deployment

Main.Gamesparks.import_game_configuration

Main.Gamesparks.list_extension_versions

Main.Gamesparks.list_extensions

Main.Gamesparks.list_games

Main.Gamesparks.list_generated_code_jobs

Main.Gamesparks.list_snapshots

Main.Gamesparks.list_stage_deployments

Main.Gamesparks.list_stages

Main.Gamesparks.list_tags_for_resource

Main.Gamesparks.start_generated_code_job

Main.Gamesparks.start_stage_deployment

Main.Gamesparks.tag_resource

Main.Gamesparks.untag_resource

Main.Gamesparks.update_game

Main.Gamesparks.update_game_configuration

Main.Gamesparks.update_snapshot

Main.Gamesparks.update_stage

Documentation

Main.Gamesparks.create_game — Method

create_game(game_name)
create_game(game_name, params::Dict{String,<:Any})

Creates a new game with an empty configuration. After creating your game, you can update the configuration using UpdateGameConfiguration or ImportGameConfiguration.

Arguments

game_name: The name of the game.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": A client-defined token. With an active client token in the request, this action is idempotent.

"Description": The description of the game.

"Tags": The list of tags to apply to the game.

Main.Gamesparks.create_snapshot — Method

create_snapshot(game_name)
create_snapshot(game_name, params::Dict{String,<:Any})

Creates a snapshot of the game configuration.

Arguments

game_name: The name of the game.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": The description of the snapshot.

Main.Gamesparks.create_stage — Method

create_stage(game_name, role, stage_name)
create_stage(game_name, role, stage_name, params::Dict{String,<:Any})

Creates a new stage for stage-by-stage game development and deployment.

Arguments

game_name: The name of the game.

role: The Amazon Resource Name (ARN) of the role to run the game with. This role can be a game-defined role or the default role that GameSparks created.

stage_name: The name of the stage.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": A client-defined token. With an active client token in the request, this action is idempotent.

"Description": The description of the stage.

"Tags": The list of tags to apply to the stage.

Main.Gamesparks.delete_game — Method

delete_game(game_name)
delete_game(game_name, params::Dict{String,<:Any})

Deletes a game.

Arguments

game_name: The name of the game to delete.

Main.Gamesparks.delete_stage — Method

delete_stage(game_name, stage_name)
delete_stage(game_name, stage_name, params::Dict{String,<:Any})

Deletes a stage from a game, along with the associated game runtime.

Arguments

game_name: The name of the game.

stage_name: The name of the stage to delete.

Main.Gamesparks.disconnect_player — Method

disconnect_player(game_name, player_id, stage_name)
disconnect_player(game_name, player_id, stage_name, params::Dict{String,<:Any})

Disconnects a player from the game runtime. If a player has multiple connections, this operation attempts to close all of them.

Arguments

game_name: The name of the game.

player_id: The unique identifier representing a player.

stage_name: The name of the stage.

Main.Gamesparks.export_snapshot — Method

export_snapshot(game_name, snapshot_id)
export_snapshot(game_name, snapshot_id, params::Dict{String,<:Any})

Exports a game configuration snapshot.

Arguments

game_name: The name of the game.

snapshot_id: The identifier of the snapshot to export.

Main.Gamesparks.get_extension — Method

get_extension(name, namespace)
get_extension(name, namespace, params::Dict{String,<:Any})

Gets details about a specified extension.

Arguments

name: The name of the extension.

namespace: The namespace (qualifier) of the extension.

Main.Gamesparks.get_extension_version — Method

get_extension_version(extension_version, name, namespace)
get_extension_version(extension_version, name, namespace, params::Dict{String,<:Any})

Gets details about a specified extension version.

Arguments

extension_version: The version of the extension.

name: The name of the extension.

namespace: The namespace (qualifier) of the extension.

Main.Gamesparks.get_game — Method

get_game(game_name)
get_game(game_name, params::Dict{String,<:Any})

Gets details about a game.

Arguments

game_name: The name of the game.

Main.Gamesparks.get_game_configuration — Method

get_game_configuration(game_name)
get_game_configuration(game_name, params::Dict{String,<:Any})

Gets the configuration of the game.

Arguments

game_name: The name of the game.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Sections": The list of sections to return.

Main.Gamesparks.get_generated_code_job — Method

get_generated_code_job(game_name, job_id, snapshot_id)
get_generated_code_job(game_name, job_id, snapshot_id, params::Dict{String,<:Any})

Gets details about a job that is generating code for a snapshot.

Arguments

game_name: The name of the game.

job_id: The identifier of the code generation job.

snapshot_id: The identifier of the snapshot for the code generation job.

Main.Gamesparks.get_player_connection_status — Method

get_player_connection_status(game_name, player_id, stage_name)
get_player_connection_status(game_name, player_id, stage_name, params::Dict{String,<:Any})

Gets the status of a player's connection to the game runtime. It's possible for a single player to have multiple connections to the game runtime. If a player is not connected, this operation returns an empty list.

Arguments

game_name: The name of the game.

player_id: The unique identifier representing a player.

stage_name: The name of the stage.

Main.Gamesparks.get_snapshot — Method

get_snapshot(game_name, snapshot_id)
get_snapshot(game_name, snapshot_id, params::Dict{String,<:Any})

Gets a copy of the game configuration in a snapshot.

Arguments

game_name: The name of the game.

snapshot_id: The identifier of the snapshot.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Sections": The list of game configuration sections to be described.

Main.Gamesparks.get_stage — Method

get_stage(game_name, stage_name)
get_stage(game_name, stage_name, params::Dict{String,<:Any})

Gets information about a stage.

Arguments

game_name: The name of the game.

stage_name: The name of the stage.

Main.Gamesparks.get_stage_deployment — Method

get_stage_deployment(game_name, stage_name)
get_stage_deployment(game_name, stage_name, params::Dict{String,<:Any})

Gets information about a stage deployment.

Arguments

game_name: The name of the game.

stage_name: The name of the stage.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DeploymentId": The identifier of the stage deployment. StartStageDeployment returns the identifier that you use here.

Main.Gamesparks.import_game_configuration — Method

import_game_configuration(game_name, import_source)
import_game_configuration(game_name, import_source, params::Dict{String,<:Any})

Imports a game configuration. This operation replaces the current configuration of the game with the provided input. This is not a reversible operation. If you want to preserve the previous configuration, use CreateSnapshot to make a new snapshot before importing.

Arguments

game_name: The name of the game.

import_source: The source used to import configuration sections.

Main.Gamesparks.list_extension_versions — Method

list_extension_versions(name, namespace)
list_extension_versions(name, namespace, params::Dict{String,<:Any})

Gets a paginated list of available versions for the extension. Each time an API change is made to an extension, the version is incremented. The list retrieved by this operation shows the versions that are currently available.

Arguments

name: The name of the extension.

namespace: The namespace (qualifier) of the extension.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

"NextToken": The token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

Main.Gamesparks.list_extensions — Method

list_extensions()
list_extensions(params::Dict{String,<:Any})

Gets a paginated list of available extensions. Extensions provide features that games can use from scripts.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

"NextToken": The token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

Main.Gamesparks.list_games — Method

list_games()
list_games(params::Dict{String,<:Any})

Gets a paginated list of games.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

"NextToken": The token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

Main.Gamesparks.list_generated_code_jobs — Method

list_generated_code_jobs(game_name, snapshot_id)
list_generated_code_jobs(game_name, snapshot_id, params::Dict{String,<:Any})

Gets a paginated list of code generation jobs for a snapshot.

Arguments

game_name: The name of the game.

snapshot_id: The identifier of the snapshot.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

"NextToken": The token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

Main.Gamesparks.list_snapshots — Method

list_snapshots(game_name)
list_snapshots(game_name, params::Dict{String,<:Any})

Gets a paginated list of snapshot summaries from the game.

Arguments

game_name: The name of the game.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

"NextToken": The token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

Main.Gamesparks.list_stage_deployments — Method

list_stage_deployments(game_name, stage_name)
list_stage_deployments(game_name, stage_name, params::Dict{String,<:Any})

Gets a paginated list of stage deployment summaries from the game.

Arguments

game_name: The name of the game.

stage_name: The name of the stage.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

"NextToken": The token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

Main.Gamesparks.list_stages — Method

list_stages(game_name)
list_stages(game_name, params::Dict{String,<:Any})

Gets a paginated list of stage summaries from the game.

Arguments

game_name: The name of the game.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return. Use this parameter with NextToken to get results as a set of sequential pages.

"NextToken": The token that indicates the start of the next sequential page of results. Use the token that is returned with a previous call to this operation. To start at the beginning of the result set, do not specify a value.

Main.Gamesparks.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists the tags associated with a GameSparks resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the GameSparks resource.

Main.Gamesparks.start_generated_code_job — Method

start_generated_code_job(game_name, generator, snapshot_id)
start_generated_code_job(game_name, generator, snapshot_id, params::Dict{String,<:Any})

Starts an asynchronous process that generates client code for system-defined and custom messages. The resulting code is collected as a .zip file and uploaded to a pre-signed Amazon S3 URL.

Arguments

game_name: The name of the game.

generator: Properties of the generator to use for the job.

snapshot_id: The identifier of the snapshot for which to generate code.

Main.Gamesparks.start_stage_deployment — Method

start_stage_deployment(game_name, snapshot_id, stage_name)
start_stage_deployment(game_name, snapshot_id, stage_name, params::Dict{String,<:Any})

Deploys a snapshot to the stage and creates a new game runtime. After you call this operation, you can check the deployment status by using GetStageDeployment. If there are any players connected to the previous game runtime, then both runtimes persist. Existing connections to the previous runtime are maintained. When players disconnect and reconnect, they connect to the new runtime. After there are no connections to the previous game runtime, it is deleted.

Arguments

game_name: The name of the game.

snapshot_id: The identifier of the snapshot to deploy.

stage_name: The name of the stage to deploy the snapshot onto.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": A client-defined token. With an active client token in the request, this action is idempotent.

Main.Gamesparks.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Adds tags to a GameSparks resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource to add the tags to.

tags: The tags to add to the resource.

Main.Gamesparks.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes tags from a GameSparks resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource to remove the tags from.

tag_keys: The keys of the tags to remove.

Main.Gamesparks.update_game — Method

update_game(game_name)
update_game(game_name, params::Dict{String,<:Any})

Updates details of the game.

Arguments

game_name: The name of the game.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": The description of the game.

Main.Gamesparks.update_game_configuration — Method

update_game_configuration(game_name, modifications)
update_game_configuration(game_name, modifications, params::Dict{String,<:Any})

Updates one or more sections of the game configuration.

Arguments

game_name: The name of the game.

modifications: The list of modifications to make.

Main.Gamesparks.update_snapshot — Method

update_snapshot(game_name, snapshot_id)
update_snapshot(game_name, snapshot_id, params::Dict{String,<:Any})

Updates the metadata of a GameSparks snapshot.

Arguments

game_name: The name of the game.

snapshot_id: The identifier of the snapshot.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": The description of the snapshot.

Main.Gamesparks.update_stage — Method

update_stage(game_name, stage_name)
update_stage(game_name, stage_name, params::Dict{String,<:Any})

Updates the metadata of a stage.

Arguments

game_name: The name of the game.

stage_name: The name of the stage.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": The description of the stage.

"Role": The Amazon Resource Name (ARN) of the role to use for the game snapshots deployed to this stage.