Geo Maps

This page documents function available when using the Geo_Maps module, created with @service Geo_Maps.

Index

Main.Geo_Maps.get_glyphs

Main.Geo_Maps.get_sprites

Main.Geo_Maps.get_static_map

Main.Geo_Maps.get_style_descriptor

Main.Geo_Maps.get_tile

Documentation

Main.Geo_Maps.get_glyphs — Function

get_glyphs(font_stack, font_unicode_range)
get_glyphs(font_stack, font_unicode_range, params::Dict{String,<:Any})

GetGlyphs returns the map's glyphs.

For more information, see Style labels with glyphs in the Amazon Location Service Developer Guide.

Arguments

font_stack: Name of the FontStack to retrieve.

Example: Amazon Ember Bold,Noto Sans Bold.

The supported font stacks are as follows:

Amazon Ember Bold

Amazon Ember Bold Italic

Amazon Ember Bold,Noto Sans Bold

Amazon Ember Bold,Noto Sans Bold,Noto Sans Arabic Bold

Amazon Ember Condensed RC BdItalic

Amazon Ember Condensed RC Bold

Amazon Ember Condensed RC Bold Italic

Amazon Ember Condensed RC Bold,Noto Sans Bold

Amazon Ember Condensed RC Bold,Noto Sans Bold,Noto Sans Arabic Condensed Bold

Amazon Ember Condensed RC Light

Amazon Ember Condensed RC Light Italic

Amazon Ember Condensed RC LtItalic

Amazon Ember Condensed RC Regular

Amazon Ember Condensed RC Regular Italic

Amazon Ember Condensed RC Regular,Noto Sans Regular

Amazon Ember Condensed RC Regular,Noto Sans Regular,Noto Sans Arabic Condensed Regular

Amazon Ember Condensed RC RgItalic

Amazon Ember Condensed RC ThItalic

Amazon Ember Condensed RC Thin

Amazon Ember Condensed RC Thin Italic

Amazon Ember Heavy

Amazon Ember Heavy Italic

Amazon Ember Light

Amazon Ember Light Italic

Amazon Ember Medium

Amazon Ember Medium Italic

Amazon Ember Medium,Noto Sans Medium

Amazon Ember Medium,Noto Sans Medium,Noto Sans Arabic Medium

Amazon Ember Regular

Amazon Ember Regular Italic

Amazon Ember Regular Italic,Noto Sans Italic

Amazon Ember Regular Italic,Noto Sans Italic,Noto Sans Arabic Regular

Amazon Ember Regular,Noto Sans Regular

Amazon Ember Regular,Noto Sans Regular,Noto Sans Arabic Regular

Amazon Ember Thin

Amazon Ember Thin Italic

AmazonEmberCdRC_Bd

AmazonEmberCdRC_BdIt

AmazonEmberCdRC_Lt

AmazonEmberCdRC_LtIt

AmazonEmberCdRC_Rg

AmazonEmberCdRC_RgIt

AmazonEmberCdRC_Th

AmazonEmberCdRC_ThIt

AmazonEmber_Bd

AmazonEmber_BdIt

AmazonEmber_He

AmazonEmber_HeIt

AmazonEmber_Lt

AmazonEmber_LtIt

AmazonEmber_Md

AmazonEmber_MdIt

AmazonEmber_Rg

AmazonEmber_RgIt

AmazonEmber_Th

AmazonEmber_ThIt

Noto Sans Black

Noto Sans Black Italic

Noto Sans Bold

Noto Sans Bold Italic

Noto Sans Extra Bold

Noto Sans Extra Bold Italic

Noto Sans Extra Light

Noto Sans Extra Light Italic

Noto Sans Italic

Noto Sans Light

Noto Sans Light Italic

Noto Sans Medium

Noto Sans Medium Italic

Noto Sans Regular

Noto Sans Semi Bold

Noto Sans Semi Bold Italic

Noto Sans Thin

Noto Sans Thin Italic

NotoSans-Bold

NotoSans-Italic

NotoSans-Medium

NotoSans-Regular

Open Sans Regular,Arial Unicode MS Regular

font_unicode_range: A Unicode range of characters to download glyphs for. This must be aligned to multiples of 256.

Example: 0-255.pbf

Main.Geo_Maps.get_sprites — Function

get_sprites(color_scheme, file_name, style, variant)
get_sprites(color_scheme, file_name, style, variant, params::Dict{String,<:Any})

GetSprites returns the map's sprites.

For more information, see Style iconography with sprites in the Amazon Location Service Developer Guide.

Arguments

color_scheme: Sets the color tone for the map sprites, such as dark and light.

Example: Light

Default value: Light

Note

Valid values for ColorScheme are case sensitive.

file_name: Sprites API: The name of the sprite ﬁle to retrieve, following pattern sprites(@2x)?\.(png|json).

Example: sprites.png

style: Style specifies the desired map style for the Sprites APIs.

variant: Optimizes map styles for specific use case or industry. You can choose allowed variant only with Standard map style.

Example: Default

Note

Valid values for Variant are case sensitive.

Main.Geo_Maps.get_static_map — Function

get_static_map(file_name, height, width)
get_static_map(file_name, height, width, params::Dict{String,<:Any})

Note

This operation is not supported in ap-southeast-1 and ap-southeast-5 regions for GrabMaps customers.

GetStaticMap provides high-quality static map images with customizable options. You can modify the map's appearance and overlay additional information. It's an ideal solution for applications requiring tailored static map snapshots.

For more information, see the following topics in the Amazon Location Service Developer Guide:

Static maps

Customize static maps

Overlay on the static map

Arguments

file_name: The map scaling parameter to size the image, icons, and labels. It follows the pattern of ^map(@2x)?$.

Example: map, map@2x

height: Specifies the height of the map image.

width: Specifies the width of the map image.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"bounded-positions": Takes in two or more pair of coordinates in World Geodetic System (WGS 84) format: [longitude, latitude], with each coordinate separated by a comma. The API will generate an image to encompass all of the provided coordinates.

Note

Cannot be used with Zoom and or Radius

Example: 97.170451,78.039098,99.045536,27.176178

"bounding-box": Takes in two pairs of coordinates in World Geodetic System (WGS 84) format: [longitude, latitude], denoting south-westerly and north-easterly edges of the image. The underlying area becomes the view of the image.

Example: -123.17075,49.26959,-123.08125,49.31429

"center": Takes in a pair of coordinates in World Geodetic System (WGS 84) format: [longitude, latitude], which becomes the center point of the image. This parameter requires that either zoom or radius is set.

Note

Cannot be used with Zoom and or Radius

Example: 49.295,-123.108

"color-scheme": Sets the color tone for the map, such as dark and light.

Example: Light

Default value: Light

Note

Valid values for ColorScheme are case sensitive.

"compact-overlay": Takes in a string to draw geometries on the image. The input is a comma separated format as follows format: [Lon, Lat]

Example: line:-122.407653,37.798557,-122.413291,37.802443;color=%23DD0000;width=7;outline-color=#00DD00;outline-width=5yd|point:-122.40572,37.80004;label=Fog Hill Market;size=large;text-color=%23DD0000;color=#EE4B2B

Note

Currently it supports the following geometry types: point, line and polygon. It does not support multiPoint , multiLine and multiPolgyon.

"crop-labels": It is a flag that takes in true or false. It prevents the labels that are on the edge of the image from being cut or obscured.

"geojson-overlay": Takes in a string to draw geometries on the image. The input is a valid GeoJSON collection object.

Example: {"type":"FeatureCollection","features": [{"type":"Feature","geometry":{"type":"MultiPoint","coordinates": [[-90.076345,51.504107],[-0.074451,51.506892]]},"properties": {"color":"#00DD00"}}]}

"key": Optional: The API key to be used for authorization. Either an API key or valid SigV4 signature must be provided when making a request.

"label-size": Overrides the label size auto-calculated by FileName. Takes in one of the values - Small or Large.

"lang": Specifies the language on the map labels using the BCP 47 language tag, limited to ISO 639-1 two-letter language codes. If the specified language data isn't available for the map image, the labels will default to the regional primary language.

Supported codes:

ar

as

az

be

bg

bn

bs

ca

cs

cy

da

de

el

en

es

et

eu

fi

fo

fr

ga

gl

gn

gu

he

hi

hr

hu

hy

id

is

it

ja

ka

kk

km

kn

ko

ky

lt

lv

mk

ml

mr

ms

mt

my

nl

no

or

pa

pl

pt

ro

ru

sk

sl

sq

sr

sv

ta

te

th

tr

uk

uz

vi

zh

"padding": Applies additional space (in pixels) around overlay feature to prevent them from being cut or obscured.

Note

Value for max and min is determined by:

Min: 1

Max: min(height, width)/4

Example: 100

"pois": Determines if the result image will display icons representing points of interest on the map.

"political-view": Specifies the political view, using ISO 3166-2 or ISO 3166-3 country code format.

The following political views are currently supported:

ARG: Argentina's view on the Southern Patagonian Ice Field and Tierra Del Fuego, including the Falkland Islands, South Georgia, and South Sandwich Islands

EGY: Egypt's view on Bir Tawil

IND: India's view on Gilgit-Baltistan

KEN: Kenya's view on the Ilemi Triangle

MAR: Morocco's view on Western Sahara

RUS: Russia's view on Crimea

SDN: Sudan's view on the Halaib Triangle

SRB: Serbia's view on Kosovo, Vukovar, and Sarengrad Islands

SUR: Suriname's view on the Courantyne Headwaters and Lawa Headwaters

SYR: Syria's view on the Golan Heights

TUR: Turkey's view on Cyprus and Northern Cyprus

TZA: Tanzania's view on Lake Malawi

URY: Uruguay's view on Rincon de Artigas

VNM: Vietnam's view on the Paracel Islands and Spratly Islands

"radius": Used with center parameter, it specifies the zoom of the image where you can control it on a granular level. Takes in any value &gt;= 1.

Example: 1500

Note

Cannot be used with Zoom.

Unit: Meters

"scale-unit": Displays a scale on the bottom right of the map image with the unit specified in the input.

Example: KilometersMiles, Miles, Kilometers, MilesKilometers

"style": Style specifies the desired map style.

"zoom": Specifies the zoom level of the map image.

Note

Cannot be used with Radius.

Main.Geo_Maps.get_style_descriptor — Function

get_style_descriptor(style)
get_style_descriptor(style, params::Dict{String,<:Any})

GetStyleDescriptor returns information about the style.

For more information, see Style dynamic maps in the Amazon Location Service Developer Guide.

Arguments

style: Style specifies the desired map style. For GrabMaps customers, ap-southeast-1 and ap-southeast-5 regions support only the Standard and Monochrome values.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"buildings": Adjusts how building details are rendered on the map.

The following building styles are currently supported:

Buildings3D: Displays buildings as three-dimensional extrusions on the map.

Buildings3D is valid only for the Standard and Monochrome map styles.

"color-scheme": Sets the color tone for the map, such as dark and light.

Example: Light

Default value: Light

Note

Valid values for ColorScheme are case sensitive.

"contour-density": Displays the shape and steepness of terrain features using elevation lines. The density value controls how densely the available contour line information is rendered on the map. Not supported in ap-southeast-1 and ap-southeast-5 regions for GrabMaps customers.

This parameter is valid for all map styles except Satellite.

"key": Optional: The API key to be used for authorization. Either an API key or valid SigV4 signature must be provided when making a request.

"political-view": Specifies the political view using ISO 3166-2 or ISO 3166-3 country code format. Not supported in ap-southeast-1 and ap-southeast-5 regions for GrabMaps customers.

The following political views are currently supported:

ARG: Argentina's view on the Southern Patagonian Ice Field and Tierra Del Fuego, including the Falkland Islands, South Georgia, and South Sandwich Islands

EGY: Egypt's view on Bir Tawil

IND: India's view on Gilgit-Baltistan

KEN: Kenya's view on the Ilemi Triangle

MAR: Morocco's view on Western Sahara

RUS: Russia's view on Crimea

SDN: Sudan's view on the Halaib Triangle

SRB: Serbia's view on Kosovo, Vukovar, and Sarengrad Islands

SUR: Suriname's view on the Courantyne Headwaters and Lawa Headwaters

SYR: Syria's view on the Golan Heights

TUR: Turkey's view on Cyprus and Northern Cyprus

TZA: Tanzania's view on Lake Malawi

URY: Uruguay's view on Rincon de Artigas

VNM: Vietnam's view on the Paracel Islands and Spratly Islands

"terrain": Adjusts how physical terrain details are rendered on the map. Not supported in ap-southeast-1 and ap-southeast-5 regions for GrabMaps customers.

The following terrain styles are currently supported:

Hillshade: Displays the physical terrain details through shading and highlighting of elevation change and geographic features.

Terrain3D: Displays physical terrain details and elevations as a three-dimensional model.

Hillshade is valid only for the Standard and Monochrome map styles.

"traffic": Displays real-time traffic information overlay on map, such as incident events and flow events. Not supported in ap-southeast-1 and ap-southeast-5 regions for GrabMaps customers.

This parameter is valid for all map styles except Satellite.

"travel-modes": Renders additional map information relevant to selected travel modes. Information for multiple travel modes can be displayed simultaneously, although this increases the overall information density rendered on the map. Not supported in ap-southeast-1 and ap-southeast-5 regions for GrabMaps customers.

This parameter is valid for all map styles except Satellite.

Main.Geo_Maps.get_tile — Function

get_tile(tileset, x, y, z)
get_tile(tileset, x, y, z, params::Dict{String,<:Any})

GetTile returns a tile. Map tiles are used by clients to render a map. They're addressed using a grid arrangement with an X coordinate, Y coordinate, and Z (zoom) level.

For more information, see Tiles in the Amazon Location Service Developer Guide.

Arguments

tileset: Specifies the desired tile set. For GrabMaps customers, ap-southeast-1 and ap-southeast-5 regions support only the vector.basemap value.

Valid Values: raster.satellite | vector.basemap | vector.traffic | raster.dem

x: The X axis value for the map tile.

y: The Y axis value for the map tile.

z: The zoom value for the map tile.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"additional-features": A list of optional additional parameters such as map styles that can be requested for each result. Not supported in ap-southeast-1 and ap-southeast-5 regions for GrabMaps customers.

"key": Optional: The API key to be used for authorization. Either an API key or valid SigV4 signature must be provided when making a request.