Global Accelerator

This page documents function available when using the Global_Accelerator module, created with @service Global_Accelerator.

Index

Main.Global_Accelerator.add_custom_routing_endpoints

Main.Global_Accelerator.advertise_byoip_cidr

Main.Global_Accelerator.allow_custom_routing_traffic

Main.Global_Accelerator.create_accelerator

Main.Global_Accelerator.create_custom_routing_accelerator

Main.Global_Accelerator.create_custom_routing_endpoint_group

Main.Global_Accelerator.create_custom_routing_listener

Main.Global_Accelerator.create_endpoint_group

Main.Global_Accelerator.create_listener

Main.Global_Accelerator.delete_accelerator

Main.Global_Accelerator.delete_custom_routing_accelerator

Main.Global_Accelerator.delete_custom_routing_endpoint_group

Main.Global_Accelerator.delete_custom_routing_listener

Main.Global_Accelerator.delete_endpoint_group

Main.Global_Accelerator.delete_listener

Main.Global_Accelerator.deny_custom_routing_traffic

Main.Global_Accelerator.deprovision_byoip_cidr

Main.Global_Accelerator.describe_accelerator

Main.Global_Accelerator.describe_accelerator_attributes

Main.Global_Accelerator.describe_custom_routing_accelerator

Main.Global_Accelerator.describe_custom_routing_accelerator_attributes

Main.Global_Accelerator.describe_custom_routing_endpoint_group

Main.Global_Accelerator.describe_custom_routing_listener

Main.Global_Accelerator.describe_endpoint_group

Main.Global_Accelerator.describe_listener

Main.Global_Accelerator.list_accelerators

Main.Global_Accelerator.list_byoip_cidrs

Main.Global_Accelerator.list_custom_routing_accelerators

Main.Global_Accelerator.list_custom_routing_endpoint_groups

Main.Global_Accelerator.list_custom_routing_listeners

Main.Global_Accelerator.list_custom_routing_port_mappings

Main.Global_Accelerator.list_custom_routing_port_mappings_by_destination

Main.Global_Accelerator.list_endpoint_groups

Main.Global_Accelerator.list_listeners

Main.Global_Accelerator.list_tags_for_resource

Main.Global_Accelerator.provision_byoip_cidr

Main.Global_Accelerator.remove_custom_routing_endpoints

Main.Global_Accelerator.tag_resource

Main.Global_Accelerator.untag_resource

Main.Global_Accelerator.update_accelerator

Main.Global_Accelerator.update_accelerator_attributes

Main.Global_Accelerator.update_custom_routing_accelerator

Main.Global_Accelerator.update_custom_routing_accelerator_attributes

Main.Global_Accelerator.update_custom_routing_listener

Main.Global_Accelerator.update_endpoint_group

Main.Global_Accelerator.update_listener

Main.Global_Accelerator.withdraw_byoip_cidr

Documentation

Main.Global_Accelerator.add_custom_routing_endpoints — Method

add_custom_routing_endpoints(endpoint_configurations, endpoint_group_arn)
add_custom_routing_endpoints(endpoint_configurations, endpoint_group_arn, params::Dict{String,<:Any})

Associate a virtual private cloud (VPC) subnet endpoint with your custom routing accelerator. The listener port range must be large enough to support the number of IP addresses that can be specified in your subnet. The number of ports required is: subnet size times the number of ports per destination EC2 instances. For example, a subnet defined as /24 requires a listener port range of at least 255 ports. Note: You must have enough remaining listener ports available to map to the subnet ports, or the call will fail with a LimitExceededException. By default, all destinations in a subnet in a custom routing accelerator cannot receive traffic. To enable all destinations to receive traffic, or to specify individual port mappings that can receive traffic, see the AllowCustomRoutingTraffic operation.

Arguments

endpoint_configurations: The list of endpoint objects to add to a custom routing accelerator.

endpoint_group_arn: The Amazon Resource Name (ARN) of the endpoint group for the custom routing endpoint.

Main.Global_Accelerator.advertise_byoip_cidr — Method

advertise_byoip_cidr(cidr)
advertise_byoip_cidr(cidr, params::Dict{String,<:Any})

Advertises an IPv4 address range that is provisioned for use with your AWS resources through bring your own IP addresses (BYOIP). It can take a few minutes before traffic to the specified addresses starts routing to AWS because of propagation delays. To stop advertising the BYOIP address range, use WithdrawByoipCidr. For more information, see Bring Your Own IP Addresses (BYOIP) in the AWS Global Accelerator Developer Guide.

Arguments

cidr: The address range, in CIDR notation. This must be the exact range that you provisioned. You can't advertise only a portion of the provisioned range.

Main.Global_Accelerator.allow_custom_routing_traffic — Method

allow_custom_routing_traffic(endpoint_group_arn, endpoint_id)
allow_custom_routing_traffic(endpoint_group_arn, endpoint_id, params::Dict{String,<:Any})

Specify the Amazon EC2 instance (destination) IP addresses and ports for a VPC subnet endpoint that can receive traffic for a custom routing accelerator. You can allow traffic to all destinations in the subnet endpoint, or allow traffic to a specified list of destination IP addresses and ports in the subnet. Note that you cannot specify IP addresses or ports outside of the range that you configured for the endpoint group. After you make changes, you can verify that the updates are complete by checking the status of your accelerator: the status changes from IN_PROGRESS to DEPLOYED.

Arguments

endpoint_group_arn: The Amazon Resource Name (ARN) of the endpoint group.

endpoint_id: An ID for the endpoint. For custom routing accelerators, this is the virtual private cloud (VPC) subnet ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AllowAllTrafficToEndpoint": Indicates whether all destination IP addresses and ports for a specified VPC subnet endpoint can receive traffic from a custom routing accelerator. The value is TRUE or FALSE. When set to TRUE, all destinations in the custom routing VPC subnet can receive traffic. Note that you cannot specify destination IP addresses and ports when the value is set to TRUE. When set to FALSE (or not specified), you must specify a list of destination IP addresses that are allowed to receive traffic. A list of ports is optional. If you don't specify a list of ports, the ports that can accept traffic is the same as the ports configured for the endpoint group. The default value is FALSE.

"DestinationAddresses": A list of specific Amazon EC2 instance IP addresses (destination addresses) in a subnet that you want to allow to receive traffic. The IP addresses must be a subset of the IP addresses that you specified for the endpoint group. DestinationAddresses is required if AllowAllTrafficToEndpoint is FALSE or is not specified.

"DestinationPorts": A list of specific Amazon EC2 instance ports (destination ports) that you want to allow to receive traffic.

Main.Global_Accelerator.create_accelerator — Method

create_accelerator(idempotency_token, name)
create_accelerator(idempotency_token, name, params::Dict{String,<:Any})

Create an accelerator. An accelerator includes one or more listeners that process inbound connections and direct traffic to one or more endpoint groups, each of which includes endpoints, such as Network Load Balancers. Global Accelerator is a global service that supports endpoints in multiple AWS Regions but you must specify the US West (Oregon) Region to create or update accelerators.

Arguments

idempotency_token: A unique, case-sensitive identifier that you provide to ensure the idempotency—that is, the uniqueness—of an accelerator.

name: The name of an accelerator. The name can have a maximum of 32 characters, must contain only alphanumeric characters or hyphens (-), and must not begin or end with a hyphen.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Enabled": Indicates whether an accelerator is enabled. The value is true or false. The default value is true. If the value is set to true, an accelerator cannot be deleted. If set to false, the accelerator can be deleted.

"IpAddressType": The value for the address type must be IPv4.

"IpAddresses": Optionally, if you've added your own IP address pool to Global Accelerator (BYOIP), you can choose IP addresses from your own pool to use for the accelerator's static IP addresses when you create an accelerator. You can specify one or two addresses, separated by a space. Do not include the /32 suffix. Only one IP address from each of your IP address ranges can be used for each accelerator. If you specify only one IP address from your IP address range, Global Accelerator assigns a second static IP address for the accelerator from the AWS IP address pool. Note that you can't update IP addresses for an existing accelerator. To change them, you must create a new accelerator with the new addresses. For more information, see Bring Your Own IP Addresses (BYOIP) in the AWS Global Accelerator Developer Guide.

"Tags": Create tags for an accelerator. For more information, see Tagging in AWS Global Accelerator in the AWS Global Accelerator Developer Guide.

Main.Global_Accelerator.create_custom_routing_accelerator — Method

create_custom_routing_accelerator(idempotency_token, name)
create_custom_routing_accelerator(idempotency_token, name, params::Dict{String,<:Any})

Create a custom routing accelerator. A custom routing accelerator directs traffic to one of possibly thousands of Amazon EC2 instance destinations running in a single or multiple virtual private clouds (VPC) subnet endpoints. Be aware that, by default, all destination EC2 instances in a VPC subnet endpoint cannot receive traffic. To enable all destinations to receive traffic, or to specify individual port mappings that can receive traffic, see the AllowCustomRoutingTraffic operation. Global Accelerator is a global service that supports endpoints in multiple AWS Regions but you must specify the US West (Oregon) Region to create or update accelerators.

Arguments

idempotency_token: A unique, case-sensitive identifier that you provide to ensure the idempotency—that is, the uniqueness—of the request.

name: The name of a custom routing accelerator. The name can have a maximum of 64 characters, must contain only alphanumeric characters or hyphens (-), and must not begin or end with a hyphen.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Enabled": Indicates whether an accelerator is enabled. The value is true or false. The default value is true. If the value is set to true, an accelerator cannot be deleted. If set to false, the accelerator can be deleted.

"IpAddressType": The value for the address type must be IPv4.

"IpAddresses": Optionally, if you've added your own IP address pool to Global Accelerator (BYOIP), you can choose IP addresses from your own pool to use for the accelerator's static IP addresses when you create an accelerator. You can specify one or two addresses, separated by a space. Do not include the /32 suffix. Only one IP address from each of your IP address ranges can be used for each accelerator. If you specify only one IP address from your IP address range, Global Accelerator assigns a second static IP address for the accelerator from the AWS IP address pool. Note that you can't update IP addresses for an existing accelerator. To change them, you must create a new accelerator with the new addresses. For more information, see Bring your own IP addresses (BYOIP) in the AWS Global Accelerator Developer Guide.

"Tags": Create tags for an accelerator. For more information, see Tagging in AWS Global Accelerator in the AWS Global Accelerator Developer Guide.

Main.Global_Accelerator.create_custom_routing_endpoint_group — Method

create_custom_routing_endpoint_group(destination_configurations, endpoint_group_region, idempotency_token, listener_arn)
create_custom_routing_endpoint_group(destination_configurations, endpoint_group_region, idempotency_token, listener_arn, params::Dict{String,<:Any})

Create an endpoint group for the specified listener for a custom routing accelerator. An endpoint group is a collection of endpoints in one AWS Region.

Arguments

destination_configurations: Sets the port range and protocol for all endpoints (virtual private cloud subnets) in a custom routing endpoint group to accept client traffic on.

endpoint_group_region: The AWS Region where the endpoint group is located. A listener can have only one endpoint group in a specific Region.

idempotency_token: A unique, case-sensitive identifier that you provide to ensure the idempotency—that is, the uniqueness—of the request.

listener_arn: The Amazon Resource Name (ARN) of the listener for a custom routing endpoint.

Main.Global_Accelerator.create_custom_routing_listener — Method

create_custom_routing_listener(accelerator_arn, idempotency_token, port_ranges)
create_custom_routing_listener(accelerator_arn, idempotency_token, port_ranges, params::Dict{String,<:Any})

Create a listener to process inbound connections from clients to a custom routing accelerator. Connections arrive to assigned static IP addresses on the port range that you specify.

Arguments

accelerator_arn: The Amazon Resource Name (ARN) of the accelerator for a custom routing listener.

idempotency_token: A unique, case-sensitive identifier that you provide to ensure the idempotency—that is, the uniqueness—of the request.

port_ranges: The port range to support for connections from clients to your accelerator. Separately, you set port ranges for endpoints. For more information, see About endpoints for custom routing accelerators.

Main.Global_Accelerator.create_endpoint_group — Method

create_endpoint_group(endpoint_group_region, idempotency_token, listener_arn)
create_endpoint_group(endpoint_group_region, idempotency_token, listener_arn, params::Dict{String,<:Any})

Create an endpoint group for the specified listener. An endpoint group is a collection of endpoints in one AWS Region. A resource must be valid and active when you add it as an endpoint.

Arguments

endpoint_group_region: The AWS Region where the endpoint group is located. A listener can have only one endpoint group in a specific Region.

idempotency_token: A unique, case-sensitive identifier that you provide to ensure the idempotency—that is, the uniqueness—of the request.

listener_arn: The Amazon Resource Name (ARN) of the listener.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"EndpointConfigurations": The list of endpoint objects.

"HealthCheckIntervalSeconds": The time—10 seconds or 30 seconds—between each health check for an endpoint. The default value is 30.

"HealthCheckPath": If the protocol is HTTP/S, then this specifies the path that is the destination for health check targets. The default value is slash (/).

"HealthCheckPort": The port that AWS Global Accelerator uses to check the health of endpoints that are part of this endpoint group. The default port is the listener port that this endpoint group is associated with. If listener port is a list of ports, Global Accelerator uses the first port in the list.

"HealthCheckProtocol": The protocol that AWS Global Accelerator uses to check the health of endpoints that are part of this endpoint group. The default value is TCP.

"PortOverrides": Override specific listener ports used to route traffic to endpoints that are part of this endpoint group. For example, you can create a port override in which the listener receives user traffic on ports 80 and 443, but your accelerator routes that traffic to ports 1080 and 1443, respectively, on the endpoints. For more information, see Port overrides in the AWS Global Accelerator Developer Guide.

"ThresholdCount": The number of consecutive health checks required to set the state of a healthy endpoint to unhealthy, or to set an unhealthy endpoint to healthy. The default value is 3.

"TrafficDialPercentage": The percentage of traffic to send to an AWS Region. Additional traffic is distributed to other endpoint groups for this listener. Use this action to increase (dial up) or decrease (dial down) traffic to a specific Region. The percentage is applied to the traffic that would otherwise have been routed to the Region based on optimal routing. The default value is 100.

Main.Global_Accelerator.create_listener — Method

create_listener(accelerator_arn, idempotency_token, port_ranges, protocol)
create_listener(accelerator_arn, idempotency_token, port_ranges, protocol, params::Dict{String,<:Any})

Create a listener to process inbound connections from clients to an accelerator. Connections arrive to assigned static IP addresses on a port, port range, or list of port ranges that you specify.

Arguments

accelerator_arn: The Amazon Resource Name (ARN) of your accelerator.

idempotency_token: A unique, case-sensitive identifier that you provide to ensure the idempotency—that is, the uniqueness—of the request.

port_ranges: The list of port ranges to support for connections from clients to your accelerator.

protocol: The protocol for connections from clients to your accelerator.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientAffinity": Client affinity lets you direct all requests from a user to the same endpoint, if you have stateful applications, regardless of the port and protocol of the client request. Client affinity gives you control over whether to always route each client to the same specific endpoint. AWS Global Accelerator uses a consistent-flow hashing algorithm to choose the optimal endpoint for a connection. If client affinity is NONE, Global Accelerator uses the "five-tuple" (5-tuple) properties—source IP address, source port, destination IP address, destination port, and protocol—to select the hash value, and then chooses the best endpoint. However, with this setting, if someone uses different ports to connect to Global Accelerator, their connections might not be always routed to the same endpoint because the hash value changes. If you want a given client to always be routed to the same endpoint, set client affinity to SOURCEIP instead. When you use the SOURCEIP setting, Global Accelerator uses the "two-tuple" (2-tuple) properties— source (client) IP address and destination IP address—to select the hash value. The default value is NONE.

Main.Global_Accelerator.delete_accelerator — Method

delete_accelerator(accelerator_arn)
delete_accelerator(accelerator_arn, params::Dict{String,<:Any})

Delete an accelerator. Before you can delete an accelerator, you must disable it and remove all dependent resources (listeners and endpoint groups). To disable the accelerator, update the accelerator to set Enabled to false. When you create an accelerator, by default, Global Accelerator provides you with a set of two static IP addresses. Alternatively, you can bring your own IP address ranges to Global Accelerator and assign IP addresses from those ranges. The IP addresses are assigned to your accelerator for as long as it exists, even if you disable the accelerator and it no longer accepts or routes traffic. However, when you delete an accelerator, you lose the static IP addresses that are assigned to the accelerator, so you can no longer route traffic by using them. As a best practice, ensure that you have permissions in place to avoid inadvertently deleting accelerators. You can use IAM policies with Global Accelerator to limit the users who have permissions to delete an accelerator. For more information, see Authentication and Access Control in the AWS Global Accelerator Developer Guide.

Arguments

accelerator_arn: The Amazon Resource Name (ARN) of an accelerator.

Main.Global_Accelerator.delete_custom_routing_accelerator — Method

delete_custom_routing_accelerator(accelerator_arn)
delete_custom_routing_accelerator(accelerator_arn, params::Dict{String,<:Any})

Delete a custom routing accelerator. Before you can delete an accelerator, you must disable it and remove all dependent resources (listeners and endpoint groups). To disable the accelerator, update the accelerator to set Enabled to false. When you create a custom routing accelerator, by default, Global Accelerator provides you with a set of two static IP addresses. The IP addresses are assigned to your accelerator for as long as it exists, even if you disable the accelerator and it no longer accepts or routes traffic. However, when you delete an accelerator, you lose the static IP addresses that are assigned to the accelerator, so you can no longer route traffic by using them. As a best practice, ensure that you have permissions in place to avoid inadvertently deleting accelerators. You can use IAM policies with Global Accelerator to limit the users who have permissions to delete an accelerator. For more information, see Authentication and Access Control in the AWS Global Accelerator Developer Guide.

Arguments

accelerator_arn: The Amazon Resource Name (ARN) of the custom routing accelerator to delete.

Main.Global_Accelerator.delete_custom_routing_endpoint_group — Method

delete_custom_routing_endpoint_group(endpoint_group_arn)
delete_custom_routing_endpoint_group(endpoint_group_arn, params::Dict{String,<:Any})

Delete an endpoint group from a listener for a custom routing accelerator.

Arguments

endpoint_group_arn: The Amazon Resource Name (ARN) of the endpoint group to delete.

Main.Global_Accelerator.delete_custom_routing_listener — Method

delete_custom_routing_listener(listener_arn)
delete_custom_routing_listener(listener_arn, params::Dict{String,<:Any})

Delete a listener for a custom routing accelerator.

Arguments

listener_arn: The Amazon Resource Name (ARN) of the listener to delete.

Main.Global_Accelerator.delete_endpoint_group — Method

delete_endpoint_group(endpoint_group_arn)
delete_endpoint_group(endpoint_group_arn, params::Dict{String,<:Any})

Delete an endpoint group from a listener.

Arguments

endpoint_group_arn: The Amazon Resource Name (ARN) of the endpoint group to delete.

Main.Global_Accelerator.delete_listener — Method

delete_listener(listener_arn)
delete_listener(listener_arn, params::Dict{String,<:Any})

Delete a listener from an accelerator.

Arguments

listener_arn: The Amazon Resource Name (ARN) of the listener.

Main.Global_Accelerator.deny_custom_routing_traffic — Method

deny_custom_routing_traffic(endpoint_group_arn, endpoint_id)
deny_custom_routing_traffic(endpoint_group_arn, endpoint_id, params::Dict{String,<:Any})

Specify the Amazon EC2 instance (destination) IP addresses and ports for a VPC subnet endpoint that cannot receive traffic for a custom routing accelerator. You can deny traffic to all destinations in the VPC endpoint, or deny traffic to a specified list of destination IP addresses and ports. Note that you cannot specify IP addresses or ports outside of the range that you configured for the endpoint group. After you make changes, you can verify that the updates are complete by checking the status of your accelerator: the status changes from IN_PROGRESS to DEPLOYED.

Arguments

endpoint_group_arn: The Amazon Resource Name (ARN) of the endpoint group.

endpoint_id: An ID for the endpoint. For custom routing accelerators, this is the virtual private cloud (VPC) subnet ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DenyAllTrafficToEndpoint": Indicates whether all destination IP addresses and ports for a specified VPC subnet endpoint cannot receive traffic from a custom routing accelerator. The value is TRUE or FALSE. When set to TRUE, no destinations in the custom routing VPC subnet can receive traffic. Note that you cannot specify destination IP addresses and ports when the value is set to TRUE. When set to FALSE (or not specified), you must specify a list of destination IP addresses that cannot receive traffic. A list of ports is optional. If you don't specify a list of ports, the ports that can accept traffic is the same as the ports configured for the endpoint group. The default value is FALSE.

"DestinationAddresses": A list of specific Amazon EC2 instance IP addresses (destination addresses) in a subnet that you want to prevent from receiving traffic. The IP addresses must be a subset of the IP addresses allowed for the VPC subnet associated with the endpoint group.

"DestinationPorts": A list of specific Amazon EC2 instance ports (destination ports) in a subnet endpoint that you want to prevent from receiving traffic.

Main.Global_Accelerator.deprovision_byoip_cidr — Method

deprovision_byoip_cidr(cidr)
deprovision_byoip_cidr(cidr, params::Dict{String,<:Any})

Releases the specified address range that you provisioned to use with your AWS resources through bring your own IP addresses (BYOIP) and deletes the corresponding address pool. Before you can release an address range, you must stop advertising it by using WithdrawByoipCidr and you must not have any accelerators that are using static IP addresses allocated from its address range. For more information, see Bring Your Own IP Addresses (BYOIP) in the AWS Global Accelerator Developer Guide.

Arguments

cidr: The address range, in CIDR notation. The prefix must be the same prefix that you specified when you provisioned the address range.

Main.Global_Accelerator.describe_accelerator — Method

describe_accelerator(accelerator_arn)
describe_accelerator(accelerator_arn, params::Dict{String,<:Any})

Describe an accelerator.

Arguments

accelerator_arn: The Amazon Resource Name (ARN) of the accelerator to describe.

Main.Global_Accelerator.describe_accelerator_attributes — Method

describe_accelerator_attributes(accelerator_arn)
describe_accelerator_attributes(accelerator_arn, params::Dict{String,<:Any})

Describe the attributes of an accelerator.

Arguments

accelerator_arn: The Amazon Resource Name (ARN) of the accelerator with the attributes that you want to describe.

Main.Global_Accelerator.describe_custom_routing_accelerator — Method

describe_custom_routing_accelerator(accelerator_arn)
describe_custom_routing_accelerator(accelerator_arn, params::Dict{String,<:Any})

Describe a custom routing accelerator.

Arguments

accelerator_arn: The Amazon Resource Name (ARN) of the accelerator to describe.

Main.Global_Accelerator.describe_custom_routing_accelerator_attributes — Method

describe_custom_routing_accelerator_attributes(accelerator_arn)
describe_custom_routing_accelerator_attributes(accelerator_arn, params::Dict{String,<:Any})

Describe the attributes of a custom routing accelerator.

Arguments

accelerator_arn: The Amazon Resource Name (ARN) of the custom routing accelerator to describe the attributes for.

Main.Global_Accelerator.describe_custom_routing_endpoint_group — Method

describe_custom_routing_endpoint_group(endpoint_group_arn)
describe_custom_routing_endpoint_group(endpoint_group_arn, params::Dict{String,<:Any})

Describe an endpoint group for a custom routing accelerator.

Arguments

endpoint_group_arn: The Amazon Resource Name (ARN) of the endpoint group to describe.

Main.Global_Accelerator.describe_custom_routing_listener — Method

describe_custom_routing_listener(listener_arn)
describe_custom_routing_listener(listener_arn, params::Dict{String,<:Any})

The description of a listener for a custom routing accelerator.

Arguments

listener_arn: The Amazon Resource Name (ARN) of the listener to describe.

Main.Global_Accelerator.describe_endpoint_group — Method

describe_endpoint_group(endpoint_group_arn)
describe_endpoint_group(endpoint_group_arn, params::Dict{String,<:Any})

Describe an endpoint group.

Arguments

endpoint_group_arn: The Amazon Resource Name (ARN) of the endpoint group to describe.

Main.Global_Accelerator.describe_listener — Method

describe_listener(listener_arn)
describe_listener(listener_arn, params::Dict{String,<:Any})

Describe a listener.

Arguments

listener_arn: The Amazon Resource Name (ARN) of the listener to describe.

Main.Global_Accelerator.list_accelerators — Method

list_accelerators()
list_accelerators(params::Dict{String,<:Any})

List the accelerators for an AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The number of Global Accelerator objects that you want to return with this call. The default value is 10.

"NextToken": The token for the next set of results. You receive this token from a previous call.

Main.Global_Accelerator.list_byoip_cidrs — Method

list_byoip_cidrs()
list_byoip_cidrs(params::Dict{String,<:Any})

Lists the IP address ranges that were specified in calls to ProvisionByoipCidr, including the current state and a history of state changes.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

"NextToken": The token for the next page of results.

Main.Global_Accelerator.list_custom_routing_accelerators — Method

list_custom_routing_accelerators()
list_custom_routing_accelerators(params::Dict{String,<:Any})

List the custom routing accelerators for an AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The number of custom routing Global Accelerator objects that you want to return with this call. The default value is 10.

"NextToken": The token for the next set of results. You receive this token from a previous call.

Main.Global_Accelerator.list_custom_routing_endpoint_groups — Method

list_custom_routing_endpoint_groups(listener_arn)
list_custom_routing_endpoint_groups(listener_arn, params::Dict{String,<:Any})

List the endpoint groups that are associated with a listener for a custom routing accelerator.

Arguments

listener_arn: The Amazon Resource Name (ARN) of the listener to list endpoint groups for.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The number of endpoint group objects that you want to return with this call. The default value is 10.

"NextToken": The token for the next set of results. You receive this token from a previous call.

Main.Global_Accelerator.list_custom_routing_listeners — Method

list_custom_routing_listeners(accelerator_arn)
list_custom_routing_listeners(accelerator_arn, params::Dict{String,<:Any})

List the listeners for a custom routing accelerator.

Arguments

accelerator_arn: The Amazon Resource Name (ARN) of the accelerator to list listeners for.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The number of listener objects that you want to return with this call. The default value is 10.

"NextToken": The token for the next set of results. You receive this token from a previous call.

Main.Global_Accelerator.list_custom_routing_port_mappings — Method

list_custom_routing_port_mappings(accelerator_arn)
list_custom_routing_port_mappings(accelerator_arn, params::Dict{String,<:Any})

Provides a complete mapping from the public accelerator IP address and port to destination EC2 instance IP addresses and ports in the virtual public cloud (VPC) subnet endpoint for a custom routing accelerator. For each subnet endpoint that you add, Global Accelerator creates a new static port mapping for the accelerator. The port mappings don't change after Global Accelerator generates them, so you can retrieve and cache the full mapping on your servers. If you remove a subnet from your accelerator, Global Accelerator removes (reclaims) the port mappings. If you add a subnet to your accelerator, Global Accelerator creates new port mappings (the existing ones don't change). If you add or remove EC2 instances in your subnet, the port mappings don't change, because the mappings are created when you add the subnet to Global Accelerator. The mappings also include a flag for each destination denoting which destination IP addresses and ports are allowed or denied traffic.

Arguments

accelerator_arn: The Amazon Resource Name (ARN) of the accelerator to list the custom routing port mappings for.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"EndpointGroupArn": The Amazon Resource Name (ARN) of the endpoint group to list the custom routing port mappings for.

"MaxResults": The number of destination port mappings that you want to return with this call. The default value is 10.

"NextToken": The token for the next set of results. You receive this token from a previous call.

Main.Global_Accelerator.list_custom_routing_port_mappings_by_destination — Method

list_custom_routing_port_mappings_by_destination(destination_address, endpoint_id)
list_custom_routing_port_mappings_by_destination(destination_address, endpoint_id, params::Dict{String,<:Any})

List the port mappings for a specific EC2 instance (destination) in a VPC subnet endpoint. The response is the mappings for one destination IP address. This is useful when your subnet endpoint has mappings that span multiple custom routing accelerators in your account, or for scenarios where you only want to list the port mappings for a specific destination instance.

Arguments

destination_address: The endpoint IP address in a virtual private cloud (VPC) subnet for which you want to receive back port mappings.

endpoint_id: The ID for the virtual private cloud (VPC) subnet.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The number of destination port mappings that you want to return with this call. The default value is 10.

"NextToken": The token for the next set of results. You receive this token from a previous call.

Main.Global_Accelerator.list_endpoint_groups — Method

list_endpoint_groups(listener_arn)
list_endpoint_groups(listener_arn, params::Dict{String,<:Any})

List the endpoint groups that are associated with a listener.

Arguments

listener_arn: The Amazon Resource Name (ARN) of the listener.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The number of endpoint group objects that you want to return with this call. The default value is 10.

"NextToken": The token for the next set of results. You receive this token from a previous call.

Main.Global_Accelerator.list_listeners — Method

list_listeners(accelerator_arn)
list_listeners(accelerator_arn, params::Dict{String,<:Any})

List the listeners for an accelerator.

Arguments

accelerator_arn: The Amazon Resource Name (ARN) of the accelerator for which you want to list listener objects.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The number of listener objects that you want to return with this call. The default value is 10.

"NextToken": The token for the next set of results. You receive this token from a previous call.

Main.Global_Accelerator.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

List all tags for an accelerator. For more information, see Tagging in AWS Global Accelerator in the AWS Global Accelerator Developer Guide.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the accelerator to list tags for. An ARN uniquely identifies an accelerator.

Main.Global_Accelerator.provision_byoip_cidr — Method

provision_byoip_cidr(cidr, cidr_authorization_context)
provision_byoip_cidr(cidr, cidr_authorization_context, params::Dict{String,<:Any})

Provisions an IP address range to use with your AWS resources through bring your own IP addresses (BYOIP) and creates a corresponding address pool. After the address range is provisioned, it is ready to be advertised using AdvertiseByoipCidr. For more information, see Bring Your Own IP Addresses (BYOIP) in the AWS Global Accelerator Developer Guide.

Arguments

cidr: The public IPv4 address range, in CIDR notation. The most specific IP prefix that you can specify is /24. The address range cannot overlap with another address range that you've brought to this or another Region.

cidr_authorization_context: A signed document that proves that you are authorized to bring the specified IP address range to Amazon using BYOIP.

Main.Global_Accelerator.remove_custom_routing_endpoints — Method

remove_custom_routing_endpoints(endpoint_group_arn, endpoint_ids)
remove_custom_routing_endpoints(endpoint_group_arn, endpoint_ids, params::Dict{String,<:Any})

Remove endpoints from a custom routing accelerator.

Arguments

endpoint_group_arn: The Amazon Resource Name (ARN) of the endpoint group to remove endpoints from.

endpoint_ids: The IDs for the endpoints. For custom routing accelerators, endpoint IDs are the virtual private cloud (VPC) subnet IDs.

Main.Global_Accelerator.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Add tags to an accelerator resource. For more information, see Tagging in AWS Global Accelerator in the AWS Global Accelerator Developer Guide.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the Global Accelerator resource to add tags to. An ARN uniquely identifies a resource.

tags: The tags to add to a resource. A tag consists of a key and a value that you define.

Main.Global_Accelerator.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Remove tags from a Global Accelerator resource. When you specify a tag key, the action removes both that key and its associated value. The operation succeeds even if you attempt to remove tags from an accelerator that was already removed. For more information, see Tagging in AWS Global Accelerator in the AWS Global Accelerator Developer Guide.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the Global Accelerator resource to remove tags from. An ARN uniquely identifies a resource.

tag_keys: The tag key pairs that you want to remove from the specified resources.

Main.Global_Accelerator.update_accelerator — Method

update_accelerator(accelerator_arn)
update_accelerator(accelerator_arn, params::Dict{String,<:Any})

Update an accelerator. Global Accelerator is a global service that supports endpoints in multiple AWS Regions but you must specify the US West (Oregon) Region to create or update accelerators.

Arguments

accelerator_arn: The Amazon Resource Name (ARN) of the accelerator to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Enabled": Indicates whether an accelerator is enabled. The value is true or false. The default value is true. If the value is set to true, the accelerator cannot be deleted. If set to false, the accelerator can be deleted.

"IpAddressType": The IP address type, which must be IPv4.

"Name": The name of the accelerator. The name can have a maximum of 32 characters, must contain only alphanumeric characters or hyphens (-), and must not begin or end with a hyphen.

Main.Global_Accelerator.update_accelerator_attributes — Method

update_accelerator_attributes(accelerator_arn)
update_accelerator_attributes(accelerator_arn, params::Dict{String,<:Any})

Update the attributes for an accelerator.

Arguments

accelerator_arn: The Amazon Resource Name (ARN) of the accelerator that you want to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"FlowLogsEnabled": Update whether flow logs are enabled. The default value is false. If the value is true, FlowLogsS3Bucket and FlowLogsS3Prefix must be specified. For more information, see Flow Logs in the AWS Global Accelerator Developer Guide.

"FlowLogsS3Bucket": The name of the Amazon S3 bucket for the flow logs. Attribute is required if FlowLogsEnabled is true. The bucket must exist and have a bucket policy that grants AWS Global Accelerator permission to write to the bucket.

"FlowLogsS3Prefix": Update the prefix for the location in the Amazon S3 bucket for the flow logs. Attribute is required if FlowLogsEnabled is true. If you don’t specify a prefix, the flow logs are stored in the root of the bucket. If you specify slash (/) for the S3 bucket prefix, the log file bucket folder structure will include a double slash (//), like the following: s3-bucketname//AWSLogs/awsaccount_id

Main.Global_Accelerator.update_custom_routing_accelerator — Method

update_custom_routing_accelerator(accelerator_arn)
update_custom_routing_accelerator(accelerator_arn, params::Dict{String,<:Any})

Update a custom routing accelerator.

Arguments

accelerator_arn: The Amazon Resource Name (ARN) of the accelerator to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Enabled": Indicates whether an accelerator is enabled. The value is true or false. The default value is true. If the value is set to true, the accelerator cannot be deleted. If set to false, the accelerator can be deleted.

"IpAddressType": The value for the address type must be IPv4.

"Name": The name of the accelerator. The name can have a maximum of 32 characters, must contain only alphanumeric characters or hyphens (-), and must not begin or end with a hyphen.

Main.Global_Accelerator.update_custom_routing_accelerator_attributes — Method

update_custom_routing_accelerator_attributes(accelerator_arn)
update_custom_routing_accelerator_attributes(accelerator_arn, params::Dict{String,<:Any})

Update the attributes for a custom routing accelerator.

Arguments

accelerator_arn: The Amazon Resource Name (ARN) of the custom routing accelerator to update attributes for.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"FlowLogsEnabled": Update whether flow logs are enabled. The default value is false. If the value is true, FlowLogsS3Bucket and FlowLogsS3Prefix must be specified. For more information, see Flow Logs in the AWS Global Accelerator Developer Guide.

"FlowLogsS3Bucket": The name of the Amazon S3 bucket for the flow logs. Attribute is required if FlowLogsEnabled is true. The bucket must exist and have a bucket policy that grants AWS Global Accelerator permission to write to the bucket.

"FlowLogsS3Prefix": Update the prefix for the location in the Amazon S3 bucket for the flow logs. Attribute is required if FlowLogsEnabled is true. If you don’t specify a prefix, the flow logs are stored in the root of the bucket. If you specify slash (/) for the S3 bucket prefix, the log file bucket folder structure will include a double slash (//), like the following: DOC-EXAMPLE-BUCKET//AWSLogs/awsaccountid

Main.Global_Accelerator.update_custom_routing_listener — Method

update_custom_routing_listener(listener_arn, port_ranges)
update_custom_routing_listener(listener_arn, port_ranges, params::Dict{String,<:Any})

Update a listener for a custom routing accelerator.

Arguments

listener_arn: The Amazon Resource Name (ARN) of the listener to update.

port_ranges: The updated port range to support for connections from clients to your accelerator. If you remove ports that are currently being used by a subnet endpoint, the call fails. Separately, you set port ranges for endpoints. For more information, see About endpoints for custom routing accelerators.

Main.Global_Accelerator.update_endpoint_group — Method

update_endpoint_group(endpoint_group_arn)
update_endpoint_group(endpoint_group_arn, params::Dict{String,<:Any})

Update an endpoint group. A resource must be valid and active when you add it as an endpoint.

Arguments

endpoint_group_arn: The Amazon Resource Name (ARN) of the endpoint group.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"EndpointConfigurations": The list of endpoint objects. A resource must be valid and active when you add it as an endpoint.

"HealthCheckIntervalSeconds": The time—10 seconds or 30 seconds—between each health check for an endpoint. The default value is 30.

"HealthCheckPath": If the protocol is HTTP/S, then this specifies the path that is the destination for health check targets. The default value is slash (/).

"HealthCheckPort": The port that AWS Global Accelerator uses to check the health of endpoints that are part of this endpoint group. The default port is the listener port that this endpoint group is associated with. If the listener port is a list of ports, Global Accelerator uses the first port in the list.

"HealthCheckProtocol": The protocol that AWS Global Accelerator uses to check the health of endpoints that are part of this endpoint group. The default value is TCP.

"PortOverrides": Override specific listener ports used to route traffic to endpoints that are part of this endpoint group. For example, you can create a port override in which the listener receives user traffic on ports 80 and 443, but your accelerator routes that traffic to ports 1080 and 1443, respectively, on the endpoints. For more information, see Port overrides in the AWS Global Accelerator Developer Guide.

"ThresholdCount": The number of consecutive health checks required to set the state of a healthy endpoint to unhealthy, or to set an unhealthy endpoint to healthy. The default value is 3.

"TrafficDialPercentage": The percentage of traffic to send to an AWS Region. Additional traffic is distributed to other endpoint groups for this listener. Use this action to increase (dial up) or decrease (dial down) traffic to a specific Region. The percentage is applied to the traffic that would otherwise have been routed to the Region based on optimal routing. The default value is 100.

Main.Global_Accelerator.update_listener — Method

update_listener(listener_arn)
update_listener(listener_arn, params::Dict{String,<:Any})

Update a listener.

Arguments

listener_arn: The Amazon Resource Name (ARN) of the listener to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientAffinity": Client affinity lets you direct all requests from a user to the same endpoint, if you have stateful applications, regardless of the port and protocol of the client request. Client affinity gives you control over whether to always route each client to the same specific endpoint. AWS Global Accelerator uses a consistent-flow hashing algorithm to choose the optimal endpoint for a connection. If client affinity is NONE, Global Accelerator uses the "five-tuple" (5-tuple) properties—source IP address, source port, destination IP address, destination port, and protocol—to select the hash value, and then chooses the best endpoint. However, with this setting, if someone uses different ports to connect to Global Accelerator, their connections might not be always routed to the same endpoint because the hash value changes. If you want a given client to always be routed to the same endpoint, set client affinity to SOURCEIP instead. When you use the SOURCEIP setting, Global Accelerator uses the "two-tuple" (2-tuple) properties— source (client) IP address and destination IP address—to select the hash value. The default value is NONE.

"PortRanges": The updated list of port ranges for the connections from clients to the accelerator.

"Protocol": The updated protocol for the connections from clients to the accelerator.

Main.Global_Accelerator.withdraw_byoip_cidr — Method

withdraw_byoip_cidr(cidr)
withdraw_byoip_cidr(cidr, params::Dict{String,<:Any})

Stops advertising an address range that is provisioned as an address pool. You can perform this operation at most once every 10 seconds, even if you specify different address ranges each time. It can take a few minutes before traffic to the specified addresses stops routing to AWS because of propagation delays. For more information, see Bring Your Own IP Addresses (BYOIP) in the AWS Global Accelerator Developer Guide.

Arguments

cidr: The address range, in CIDR notation.