Greengrass

This page documents function available when using the Greengrass module, created with @service Greengrass.

Index

Main.Greengrass.associate_role_to_group

Main.Greengrass.associate_service_role_to_account

Main.Greengrass.create_connector_definition

Main.Greengrass.create_connector_definition_version

Main.Greengrass.create_core_definition

Main.Greengrass.create_core_definition_version

Main.Greengrass.create_deployment

Main.Greengrass.create_device_definition

Main.Greengrass.create_device_definition_version

Main.Greengrass.create_function_definition

Main.Greengrass.create_function_definition_version

Main.Greengrass.create_group

Main.Greengrass.create_group_certificate_authority

Main.Greengrass.create_group_version

Main.Greengrass.create_logger_definition

Main.Greengrass.create_logger_definition_version

Main.Greengrass.create_resource_definition

Main.Greengrass.create_resource_definition_version

Main.Greengrass.create_software_update_job

Main.Greengrass.create_subscription_definition

Main.Greengrass.create_subscription_definition_version

Main.Greengrass.delete_connector_definition

Main.Greengrass.delete_core_definition

Main.Greengrass.delete_device_definition

Main.Greengrass.delete_function_definition

Main.Greengrass.delete_group

Main.Greengrass.delete_logger_definition

Main.Greengrass.delete_resource_definition

Main.Greengrass.delete_subscription_definition

Main.Greengrass.disassociate_role_from_group

Main.Greengrass.disassociate_service_role_from_account

Main.Greengrass.get_associated_role

Main.Greengrass.get_bulk_deployment_status

Main.Greengrass.get_connectivity_info

Main.Greengrass.get_connector_definition

Main.Greengrass.get_connector_definition_version

Main.Greengrass.get_core_definition

Main.Greengrass.get_core_definition_version

Main.Greengrass.get_deployment_status

Main.Greengrass.get_device_definition

Main.Greengrass.get_device_definition_version

Main.Greengrass.get_function_definition

Main.Greengrass.get_function_definition_version

Main.Greengrass.get_group

Main.Greengrass.get_group_certificate_authority

Main.Greengrass.get_group_certificate_configuration

Main.Greengrass.get_group_version

Main.Greengrass.get_logger_definition

Main.Greengrass.get_logger_definition_version

Main.Greengrass.get_resource_definition

Main.Greengrass.get_resource_definition_version

Main.Greengrass.get_service_role_for_account

Main.Greengrass.get_subscription_definition

Main.Greengrass.get_subscription_definition_version

Main.Greengrass.get_thing_runtime_configuration

Main.Greengrass.list_bulk_deployment_detailed_reports

Main.Greengrass.list_bulk_deployments

Main.Greengrass.list_connector_definition_versions

Main.Greengrass.list_connector_definitions

Main.Greengrass.list_core_definition_versions

Main.Greengrass.list_core_definitions

Main.Greengrass.list_deployments

Main.Greengrass.list_device_definition_versions

Main.Greengrass.list_device_definitions

Main.Greengrass.list_function_definition_versions

Main.Greengrass.list_function_definitions

Main.Greengrass.list_group_certificate_authorities

Main.Greengrass.list_group_versions

Main.Greengrass.list_groups

Main.Greengrass.list_logger_definition_versions

Main.Greengrass.list_logger_definitions

Main.Greengrass.list_resource_definition_versions

Main.Greengrass.list_resource_definitions

Main.Greengrass.list_subscription_definition_versions

Main.Greengrass.list_subscription_definitions

Main.Greengrass.list_tags_for_resource

Main.Greengrass.reset_deployments

Main.Greengrass.start_bulk_deployment

Main.Greengrass.stop_bulk_deployment

Main.Greengrass.tag_resource

Main.Greengrass.untag_resource

Main.Greengrass.update_connectivity_info

Main.Greengrass.update_connector_definition

Main.Greengrass.update_core_definition

Main.Greengrass.update_device_definition

Main.Greengrass.update_function_definition

Main.Greengrass.update_group

Main.Greengrass.update_group_certificate_configuration

Main.Greengrass.update_logger_definition

Main.Greengrass.update_resource_definition

Main.Greengrass.update_subscription_definition

Main.Greengrass.update_thing_runtime_configuration

Documentation

Main.Greengrass.associate_role_to_group — Method

AssociateRoleToGroup()

Associates a role with a group. Your Greengrass core will use the role to access AWS cloud services. The role's permissions should allow Greengrass core Lambda functions to perform actions against the cloud.

Required Parameters

GroupId: The ID of the Greengrass group.

RoleArn: The ARN of the role you wish to associate with this group. The existence of the role is not validated.

Main.Greengrass.associate_service_role_to_account — Method

AssociateServiceRoleToAccount()

Associates a role with your account. AWS IoT Greengrass will use the role to access your Lambda functions and AWS IoT resources. This is necessary for deployments to succeed. The role must have at least minimum permissions in the policy ''AWSGreengrassResourceAccessRolePolicy''.

Required Parameters

RoleArn: The ARN of the service role you wish to associate with your account.

Main.Greengrass.create_connector_definition — Method

CreateConnectorDefinition()

Creates a connector definition. You may provide the initial version of the connector definition now or use ''CreateConnectorDefinitionVersion'' at a later time.

Optional Parameters

InitialVersion: Information about the initial version of the connector definition.

Name: The name of the connector definition.

X-Amzn-Client-Token: A client token used to correlate requests and responses.

tags: Tag(s) to add to the new resource.

Main.Greengrass.create_connector_definition_version — Method

CreateConnectorDefinitionVersion()

Creates a version of a connector definition which has already been defined.

Required Parameters

ConnectorDefinitionId: The ID of the connector definition.

Optional Parameters

Connectors: A list of references to connectors in this version, with their corresponding configuration settings.

X-Amzn-Client-Token: A client token used to correlate requests and responses.

Main.Greengrass.create_core_definition — Method

CreateCoreDefinition()

Creates a core definition. You may provide the initial version of the core definition now or use ''CreateCoreDefinitionVersion'' at a later time. Greengrass groups must each contain exactly one Greengrass core.

Optional Parameters

InitialVersion: Information about the initial version of the core definition.

Name: The name of the core definition.

X-Amzn-Client-Token: A client token used to correlate requests and responses.

tags: Tag(s) to add to the new resource.

Main.Greengrass.create_core_definition_version — Method

CreateCoreDefinitionVersion()

Creates a version of a core definition that has already been defined. Greengrass groups must each contain exactly one Greengrass core.

Required Parameters

CoreDefinitionId: The ID of the core definition.

Optional Parameters

Cores: A list of cores in the core definition version.

X-Amzn-Client-Token: A client token used to correlate requests and responses.

Main.Greengrass.create_deployment — Method

CreateDeployment()

Creates a deployment. ''CreateDeployment'' requests are idempotent with respect to the ''X-Amzn-Client-Token'' token and the request parameters.

Required Parameters

DeploymentType: The type of deployment. When used for ''CreateDeployment'', only ''NewDeployment'' and ''Redeployment'' are valid.

GroupId: The ID of the Greengrass group.

Optional Parameters

DeploymentId: The ID of the deployment if you wish to redeploy a previous deployment.

GroupVersionId: The ID of the group version to be deployed.

X-Amzn-Client-Token: A client token used to correlate requests and responses.

Main.Greengrass.create_device_definition — Method

CreateDeviceDefinition()

Creates a device definition. You may provide the initial version of the device definition now or use ''CreateDeviceDefinitionVersion'' at a later time.

Optional Parameters

InitialVersion: Information about the initial version of the device definition.

Name: The name of the device definition.

X-Amzn-Client-Token: A client token used to correlate requests and responses.

tags: Tag(s) to add to the new resource.

Main.Greengrass.create_device_definition_version — Method

CreateDeviceDefinitionVersion()

Creates a version of a device definition that has already been defined.

Required Parameters

DeviceDefinitionId: The ID of the device definition.

Optional Parameters

Devices: A list of devices in the definition version.

X-Amzn-Client-Token: A client token used to correlate requests and responses.

Main.Greengrass.create_function_definition — Method

CreateFunctionDefinition()

Creates a Lambda function definition which contains a list of Lambda functions and their configurations to be used in a group. You can create an initial version of the definition by providing a list of Lambda functions and their configurations now, or use ''CreateFunctionDefinitionVersion'' later.

Optional Parameters

InitialVersion: Information about the initial version of the function definition.

Name: The name of the function definition.

X-Amzn-Client-Token: A client token used to correlate requests and responses.

tags: Tag(s) to add to the new resource.

Main.Greengrass.create_function_definition_version — Method

CreateFunctionDefinitionVersion()

Creates a version of a Lambda function definition that has already been defined.

Required Parameters

FunctionDefinitionId: The ID of the Lambda function definition.

Optional Parameters

DefaultConfig: The default configuration that applies to all Lambda functions in this function definition version. Individual Lambda functions can override these settings.

Functions: A list of Lambda functions in this function definition version.

X-Amzn-Client-Token: A client token used to correlate requests and responses.

Main.Greengrass.create_group — Method

CreateGroup()

Creates a group. You may provide the initial version of the group or use ''CreateGroupVersion'' at a later time. Tip: You can use the ''gggroupsetup'' package (https://github.com/awslabs/aws-greengrass-group-setup) as a library or command-line application to create and deploy Greengrass groups.

Optional Parameters

InitialVersion: Information about the initial version of the group.

Name: The name of the group.

X-Amzn-Client-Token: A client token used to correlate requests and responses.

tags: Tag(s) to add to the new resource.

Main.Greengrass.create_group_certificate_authority — Method

CreateGroupCertificateAuthority()

Creates a CA for the group. If a CA already exists, it will rotate the existing CA.

Required Parameters

GroupId: The ID of the Greengrass group.

Optional Parameters

X-Amzn-Client-Token: A client token used to correlate requests and responses.

Main.Greengrass.create_group_version — Method

CreateGroupVersion()

Creates a version of a group which has already been defined.

Required Parameters

GroupId: The ID of the Greengrass group.

Optional Parameters

ConnectorDefinitionVersionArn: The ARN of the connector definition version for this group.

CoreDefinitionVersionArn: The ARN of the core definition version for this group.

DeviceDefinitionVersionArn: The ARN of the device definition version for this group.

FunctionDefinitionVersionArn: The ARN of the function definition version for this group.

LoggerDefinitionVersionArn: The ARN of the logger definition version for this group.

ResourceDefinitionVersionArn: The ARN of the resource definition version for this group.

SubscriptionDefinitionVersionArn: The ARN of the subscription definition version for this group.

X-Amzn-Client-Token: A client token used to correlate requests and responses.

Main.Greengrass.create_logger_definition — Method

CreateLoggerDefinition()

Creates a logger definition. You may provide the initial version of the logger definition now or use ''CreateLoggerDefinitionVersion'' at a later time.

Optional Parameters

InitialVersion: Information about the initial version of the logger definition.

Name: The name of the logger definition.

X-Amzn-Client-Token: A client token used to correlate requests and responses.

tags: Tag(s) to add to the new resource.

Main.Greengrass.create_logger_definition_version — Method

CreateLoggerDefinitionVersion()

Creates a version of a logger definition that has already been defined.

Required Parameters

LoggerDefinitionId: The ID of the logger definition.

Optional Parameters

Loggers: A list of loggers.

X-Amzn-Client-Token: A client token used to correlate requests and responses.

Main.Greengrass.create_resource_definition — Method

CreateResourceDefinition()

Creates a resource definition which contains a list of resources to be used in a group. You can create an initial version of the definition by providing a list of resources now, or use ''CreateResourceDefinitionVersion'' later.

Optional Parameters

InitialVersion: Information about the initial version of the resource definition.

Name: The name of the resource definition.

X-Amzn-Client-Token: A client token used to correlate requests and responses.

tags: Tag(s) to add to the new resource.

Main.Greengrass.create_resource_definition_version — Method

CreateResourceDefinitionVersion()

Creates a version of a resource definition that has already been defined.

Required Parameters

ResourceDefinitionId: The ID of the resource definition.

Optional Parameters

Resources: A list of resources.

X-Amzn-Client-Token: A client token used to correlate requests and responses.

Main.Greengrass.create_software_update_job — Method

CreateSoftwareUpdateJob()

Creates a software update for a core or group of cores (specified as an IoT thing group.) Use this to update the OTA Agent as well as the Greengrass core software. It makes use of the IoT Jobs feature which provides additional commands to manage a Greengrass core software update job.

Required Parameters

S3UrlSignerRole:

SoftwareToUpdate:

UpdateTargets:

UpdateTargetsArchitecture:

UpdateTargetsOperatingSystem:

Optional Parameters

UpdateAgentLogLevel:

X-Amzn-Client-Token: A client token used to correlate requests and responses.

Main.Greengrass.create_subscription_definition — Method

CreateSubscriptionDefinition()

Creates a subscription definition. You may provide the initial version of the subscription definition now or use ''CreateSubscriptionDefinitionVersion'' at a later time.

Optional Parameters

InitialVersion: Information about the initial version of the subscription definition.

Name: The name of the subscription definition.

X-Amzn-Client-Token: A client token used to correlate requests and responses.

tags: Tag(s) to add to the new resource.

Main.Greengrass.create_subscription_definition_version — Method

CreateSubscriptionDefinitionVersion()

Creates a version of a subscription definition which has already been defined.

Required Parameters

SubscriptionDefinitionId: The ID of the subscription definition.

Optional Parameters

Subscriptions: A list of subscriptions.

X-Amzn-Client-Token: A client token used to correlate requests and responses.

Main.Greengrass.delete_connector_definition — Method

DeleteConnectorDefinition()

Deletes a connector definition.

Required Parameters

ConnectorDefinitionId: The ID of the connector definition.

Main.Greengrass.delete_core_definition — Method

DeleteCoreDefinition()

Deletes a core definition.

Required Parameters

CoreDefinitionId: The ID of the core definition.

Main.Greengrass.delete_device_definition — Method

DeleteDeviceDefinition()

Deletes a device definition.

Required Parameters

DeviceDefinitionId: The ID of the device definition.

Main.Greengrass.delete_function_definition — Method

DeleteFunctionDefinition()

Deletes a Lambda function definition.

Required Parameters

FunctionDefinitionId: The ID of the Lambda function definition.

Main.Greengrass.delete_group — Method

DeleteGroup()

Deletes a group.

Required Parameters

GroupId: The ID of the Greengrass group.

Main.Greengrass.delete_logger_definition — Method

DeleteLoggerDefinition()

Deletes a logger definition.

Required Parameters

LoggerDefinitionId: The ID of the logger definition.

Main.Greengrass.delete_resource_definition — Method

DeleteResourceDefinition()

Deletes a resource definition.

Required Parameters

ResourceDefinitionId: The ID of the resource definition.

Main.Greengrass.delete_subscription_definition — Method

DeleteSubscriptionDefinition()

Deletes a subscription definition.

Required Parameters

SubscriptionDefinitionId: The ID of the subscription definition.

Main.Greengrass.disassociate_role_from_group — Method

DisassociateRoleFromGroup()

Disassociates the role from a group.

Required Parameters

GroupId: The ID of the Greengrass group.

Main.Greengrass.disassociate_service_role_from_account — Method

DisassociateServiceRoleFromAccount()

Disassociates the service role from your account. Without a service role, deployments will not work.

Main.Greengrass.get_associated_role — Method

GetAssociatedRole()

Retrieves the role associated with a particular group.

Required Parameters

GroupId: The ID of the Greengrass group.

Main.Greengrass.get_bulk_deployment_status — Method

GetBulkDeploymentStatus()

Returns the status of a bulk deployment.

Required Parameters

BulkDeploymentId: The ID of the bulk deployment.

Main.Greengrass.get_connectivity_info — Method

GetConnectivityInfo()

Retrieves the connectivity information for a core.

Required Parameters

ThingName: The thing name.

Main.Greengrass.get_connector_definition — Method

GetConnectorDefinition()

Retrieves information about a connector definition.

Required Parameters

ConnectorDefinitionId: The ID of the connector definition.

Main.Greengrass.get_connector_definition_version — Method

GetConnectorDefinitionVersion()

Retrieves information about a connector definition version, including the connectors that the version contains. Connectors are prebuilt modules that interact with local infrastructure, device protocols, AWS, and other cloud services.

Required Parameters

ConnectorDefinitionId: The ID of the connector definition.

ConnectorDefinitionVersionId: The ID of the connector definition version. This value maps to the ''Version'' property of the corresponding ''VersionInformation'' object, which is returned by ''ListConnectorDefinitionVersions'' requests. If the version is the last one that was associated with a connector definition, the value also maps to the ''LatestVersion'' property of the corresponding ''DefinitionInformation'' object.

Optional Parameters

NextToken: The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.get_core_definition — Method

GetCoreDefinition()

Retrieves information about a core definition version.

Required Parameters

CoreDefinitionId: The ID of the core definition.

Main.Greengrass.get_core_definition_version — Method

GetCoreDefinitionVersion()

Retrieves information about a core definition version.

Required Parameters

CoreDefinitionId: The ID of the core definition.

CoreDefinitionVersionId: The ID of the core definition version. This value maps to the ''Version'' property of the corresponding ''VersionInformation'' object, which is returned by ''ListCoreDefinitionVersions'' requests. If the version is the last one that was associated with a core definition, the value also maps to the ''LatestVersion'' property of the corresponding ''DefinitionInformation'' object.

Main.Greengrass.get_deployment_status — Method

GetDeploymentStatus()

Returns the status of a deployment.

Required Parameters

DeploymentId: The ID of the deployment.

GroupId: The ID of the Greengrass group.

Main.Greengrass.get_device_definition — Method

GetDeviceDefinition()

Retrieves information about a device definition.

Required Parameters

DeviceDefinitionId: The ID of the device definition.

Main.Greengrass.get_device_definition_version — Method

GetDeviceDefinitionVersion()

Retrieves information about a device definition version.

Required Parameters

DeviceDefinitionId: The ID of the device definition.

DeviceDefinitionVersionId: The ID of the device definition version. This value maps to the ''Version'' property of the corresponding ''VersionInformation'' object, which is returned by ''ListDeviceDefinitionVersions'' requests. If the version is the last one that was associated with a device definition, the value also maps to the ''LatestVersion'' property of the corresponding ''DefinitionInformation'' object.

Optional Parameters

NextToken: The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.get_function_definition — Method

GetFunctionDefinition()

Retrieves information about a Lambda function definition, including its creation time and latest version.

Required Parameters

FunctionDefinitionId: The ID of the Lambda function definition.

Main.Greengrass.get_function_definition_version — Method

GetFunctionDefinitionVersion()

Retrieves information about a Lambda function definition version, including which Lambda functions are included in the version and their configurations.

Required Parameters

FunctionDefinitionId: The ID of the Lambda function definition.

FunctionDefinitionVersionId: The ID of the function definition version. This value maps to the ''Version'' property of the corresponding ''VersionInformation'' object, which is returned by ''ListFunctionDefinitionVersions'' requests. If the version is the last one that was associated with a function definition, the value also maps to the ''LatestVersion'' property of the corresponding ''DefinitionInformation'' object.

Optional Parameters

NextToken: The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.get_group — Method

GetGroup()

Retrieves information about a group.

Required Parameters

GroupId: The ID of the Greengrass group.

Main.Greengrass.get_group_certificate_authority — Method

GetGroupCertificateAuthority()

Retreives the CA associated with a group. Returns the public key of the CA.

Required Parameters

CertificateAuthorityId: The ID of the certificate authority.

GroupId: The ID of the Greengrass group.

Main.Greengrass.get_group_certificate_configuration — Method

GetGroupCertificateConfiguration()

Retrieves the current configuration for the CA used by the group.

Required Parameters

GroupId: The ID of the Greengrass group.

Main.Greengrass.get_group_version — Method

GetGroupVersion()

Retrieves information about a group version.

Required Parameters

GroupId: The ID of the Greengrass group.

GroupVersionId: The ID of the group version. This value maps to the ''Version'' property of the corresponding ''VersionInformation'' object, which is returned by ''ListGroupVersions'' requests. If the version is the last one that was associated with a group, the value also maps to the ''LatestVersion'' property of the corresponding ''GroupInformation'' object.

Main.Greengrass.get_logger_definition — Method

GetLoggerDefinition()

Retrieves information about a logger definition.

Required Parameters

LoggerDefinitionId: The ID of the logger definition.

Main.Greengrass.get_logger_definition_version — Method

GetLoggerDefinitionVersion()

Retrieves information about a logger definition version.

Required Parameters

LoggerDefinitionId: The ID of the logger definition.

LoggerDefinitionVersionId: The ID of the logger definition version. This value maps to the ''Version'' property of the corresponding ''VersionInformation'' object, which is returned by ''ListLoggerDefinitionVersions'' requests. If the version is the last one that was associated with a logger definition, the value also maps to the ''LatestVersion'' property of the corresponding ''DefinitionInformation'' object.

Optional Parameters

NextToken: The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.get_resource_definition — Method

GetResourceDefinition()

Retrieves information about a resource definition, including its creation time and latest version.

Required Parameters

ResourceDefinitionId: The ID of the resource definition.

Main.Greengrass.get_resource_definition_version — Method

GetResourceDefinitionVersion()

Retrieves information about a resource definition version, including which resources are included in the version.

Required Parameters

ResourceDefinitionId: The ID of the resource definition.

ResourceDefinitionVersionId: The ID of the resource definition version. This value maps to the ''Version'' property of the corresponding ''VersionInformation'' object, which is returned by ''ListResourceDefinitionVersions'' requests. If the version is the last one that was associated with a resource definition, the value also maps to the ''LatestVersion'' property of the corresponding ''DefinitionInformation'' object.

Main.Greengrass.get_service_role_for_account — Method

GetServiceRoleForAccount()

Retrieves the service role that is attached to your account.

Main.Greengrass.get_subscription_definition — Method

GetSubscriptionDefinition()

Retrieves information about a subscription definition.

Required Parameters

SubscriptionDefinitionId: The ID of the subscription definition.

Main.Greengrass.get_subscription_definition_version — Method

GetSubscriptionDefinitionVersion()

Retrieves information about a subscription definition version.

Required Parameters

SubscriptionDefinitionId: The ID of the subscription definition.

SubscriptionDefinitionVersionId: The ID of the subscription definition version. This value maps to the ''Version'' property of the corresponding ''VersionInformation'' object, which is returned by ''ListSubscriptionDefinitionVersions'' requests. If the version is the last one that was associated with a subscription definition, the value also maps to the ''LatestVersion'' property of the corresponding ''DefinitionInformation'' object.

Optional Parameters

NextToken: The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.get_thing_runtime_configuration — Method

GetThingRuntimeConfiguration()

Get the runtime configuration of a thing.

Required Parameters

ThingName: The thing name.

Main.Greengrass.list_bulk_deployment_detailed_reports — Method

ListBulkDeploymentDetailedReports()

Gets a paginated list of the deployments that have been started in a bulk deployment operation, and their current deployment status.

Required Parameters

BulkDeploymentId: The ID of the bulk deployment.

Optional Parameters

MaxResults: The maximum number of results to be returned per request.

NextToken: The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_bulk_deployments — Method

ListBulkDeployments()

Returns a list of bulk deployments.

Optional Parameters

MaxResults: The maximum number of results to be returned per request.

NextToken: The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_connector_definition_versions — Method

ListConnectorDefinitionVersions()

Lists the versions of a connector definition, which are containers for connectors. Connectors run on the Greengrass core and contain built-in integration with local infrastructure, device protocols, AWS, and other cloud services.

Required Parameters

ConnectorDefinitionId: The ID of the connector definition.

Optional Parameters

MaxResults: The maximum number of results to be returned per request.

NextToken: The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_connector_definitions — Method

ListConnectorDefinitions()

Retrieves a list of connector definitions.

Optional Parameters

MaxResults: The maximum number of results to be returned per request.

NextToken: The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_core_definition_versions — Method

ListCoreDefinitionVersions()

Lists the versions of a core definition.

Required Parameters

CoreDefinitionId: The ID of the core definition.

Optional Parameters

MaxResults: The maximum number of results to be returned per request.

NextToken: The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_core_definitions — Method

ListCoreDefinitions()

Retrieves a list of core definitions.

Optional Parameters

MaxResults: The maximum number of results to be returned per request.

NextToken: The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_deployments — Method

ListDeployments()

Returns a history of deployments for the group.

Required Parameters

GroupId: The ID of the Greengrass group.

Optional Parameters

MaxResults: The maximum number of results to be returned per request.

NextToken: The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_device_definition_versions — Method

ListDeviceDefinitionVersions()

Lists the versions of a device definition.

Required Parameters

DeviceDefinitionId: The ID of the device definition.

Optional Parameters

MaxResults: The maximum number of results to be returned per request.

NextToken: The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_device_definitions — Method

ListDeviceDefinitions()

Retrieves a list of device definitions.

Optional Parameters

MaxResults: The maximum number of results to be returned per request.

NextToken: The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_function_definition_versions — Method

ListFunctionDefinitionVersions()

Lists the versions of a Lambda function definition.

Required Parameters

FunctionDefinitionId: The ID of the Lambda function definition.

Optional Parameters

MaxResults: The maximum number of results to be returned per request.

NextToken: The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_function_definitions — Method

ListFunctionDefinitions()

Retrieves a list of Lambda function definitions.

Optional Parameters

MaxResults: The maximum number of results to be returned per request.

NextToken: The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_group_certificate_authorities — Method

ListGroupCertificateAuthorities()

Retrieves the current CAs for a group.

Required Parameters

GroupId: The ID of the Greengrass group.

Main.Greengrass.list_group_versions — Method

ListGroupVersions()

Lists the versions of a group.

Required Parameters

GroupId: The ID of the Greengrass group.

Optional Parameters

MaxResults: The maximum number of results to be returned per request.

NextToken: The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_groups — Method

ListGroups()

Retrieves a list of groups.

Optional Parameters

MaxResults: The maximum number of results to be returned per request.

NextToken: The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_logger_definition_versions — Method

ListLoggerDefinitionVersions()

Lists the versions of a logger definition.

Required Parameters

LoggerDefinitionId: The ID of the logger definition.

Optional Parameters

MaxResults: The maximum number of results to be returned per request.

NextToken: The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_logger_definitions — Method

ListLoggerDefinitions()

Retrieves a list of logger definitions.

Optional Parameters

MaxResults: The maximum number of results to be returned per request.

NextToken: The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_resource_definition_versions — Method

ListResourceDefinitionVersions()

Lists the versions of a resource definition.

Required Parameters

ResourceDefinitionId: The ID of the resource definition.

Optional Parameters

MaxResults: The maximum number of results to be returned per request.

NextToken: The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_resource_definitions — Method

ListResourceDefinitions()

Retrieves a list of resource definitions.

Optional Parameters

MaxResults: The maximum number of results to be returned per request.

NextToken: The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_subscription_definition_versions — Method

ListSubscriptionDefinitionVersions()

Lists the versions of a subscription definition.

Required Parameters

SubscriptionDefinitionId: The ID of the subscription definition.

Optional Parameters

MaxResults: The maximum number of results to be returned per request.

NextToken: The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_subscription_definitions — Method

ListSubscriptionDefinitions()

Retrieves a list of subscription definitions.

Optional Parameters

MaxResults: The maximum number of results to be returned per request.

NextToken: The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_tags_for_resource — Method

ListTagsForResource()

Retrieves a list of resource tags for a resource arn.

Required Parameters

resource-arn: The Amazon Resource Name (ARN) of the resource.

Main.Greengrass.reset_deployments — Method

ResetDeployments()

Resets a group's deployments.

Required Parameters

GroupId: The ID of the Greengrass group.

Optional Parameters

Force: If true, performs a best-effort only core reset.

X-Amzn-Client-Token: A client token used to correlate requests and responses.

Main.Greengrass.start_bulk_deployment — Method

StartBulkDeployment()

Deploys multiple groups in one operation. This action starts the bulk deployment of a specified set of group versions. Each group version deployment will be triggered with an adaptive rate that has a fixed upper limit. We recommend that you include an ''X-Amzn-Client-Token'' token in every ''StartBulkDeployment'' request. These requests are idempotent with respect to the token and the request parameters.

Required Parameters

ExecutionRoleArn: The ARN of the execution role to associate with the bulk deployment operation. This IAM role must allow the ''greengrass:CreateDeployment'' action for all group versions that are listed in the input file. This IAM role must have access to the S3 bucket containing the input file.

InputFileUri: The URI of the input file contained in the S3 bucket. The execution role must have ''getObject'' permissions on this bucket to access the input file. The input file is a JSON-serialized, line delimited file with UTF-8 encoding that provides a list of group and version IDs and the deployment type. This file must be less than 100 MB. Currently, AWS IoT Greengrass supports only ''NewDeployment'' deployment types.

Optional Parameters

X-Amzn-Client-Token: A client token used to correlate requests and responses.

tags: Tag(s) to add to the new resource.

Main.Greengrass.stop_bulk_deployment — Method

StopBulkDeployment()

Stops the execution of a bulk deployment. This action returns a status of ''Stopping'' until the deployment is stopped. You cannot start a new bulk deployment while a previous deployment is in the ''Stopping'' state. This action doesn't rollback completed deployments or cancel pending deployments.

Required Parameters

BulkDeploymentId: The ID of the bulk deployment.

Main.Greengrass.tag_resource — Method

TagResource()

Adds tags to a Greengrass resource. Valid resources are 'Group', 'ConnectorDefinition', 'CoreDefinition', 'DeviceDefinition', 'FunctionDefinition', 'LoggerDefinition', 'SubscriptionDefinition', 'ResourceDefinition', and 'BulkDeployment'.

Required Parameters

resource-arn: The Amazon Resource Name (ARN) of the resource.

Optional Parameters

tags:

Main.Greengrass.untag_resource — Method

UntagResource()

Remove resource tags from a Greengrass Resource.

Required Parameters

resource-arn: The Amazon Resource Name (ARN) of the resource.

tagKeys: An array of tag keys to delete

Main.Greengrass.update_connectivity_info — Method

UpdateConnectivityInfo()

Updates the connectivity information for the core. Any devices that belong to the group which has this core will receive this information in order to find the location of the core and connect to it.

Required Parameters

ThingName: The thing name.

Optional Parameters

ConnectivityInfo: A list of connectivity info.

Main.Greengrass.update_connector_definition — Method

UpdateConnectorDefinition()

Updates a connector definition.

Required Parameters

ConnectorDefinitionId: The ID of the connector definition.

Optional Parameters

Name: The name of the definition.

Main.Greengrass.update_core_definition — Method

UpdateCoreDefinition()

Updates a core definition.

Required Parameters

CoreDefinitionId: The ID of the core definition.

Optional Parameters

Name: The name of the definition.

Main.Greengrass.update_device_definition — Method

UpdateDeviceDefinition()

Updates a device definition.

Required Parameters

DeviceDefinitionId: The ID of the device definition.

Optional Parameters

Name: The name of the definition.

Main.Greengrass.update_function_definition — Method

UpdateFunctionDefinition()

Updates a Lambda function definition.

Required Parameters

FunctionDefinitionId: The ID of the Lambda function definition.

Optional Parameters

Name: The name of the definition.

Main.Greengrass.update_group — Method

UpdateGroup()

Updates a group.

Required Parameters

GroupId: The ID of the Greengrass group.

Optional Parameters

Name: The name of the definition.

Main.Greengrass.update_group_certificate_configuration — Method

UpdateGroupCertificateConfiguration()

Updates the Certificate expiry time for a group.

Required Parameters

GroupId: The ID of the Greengrass group.

Optional Parameters

CertificateExpiryInMilliseconds: The amount of time remaining before the certificate expires, in milliseconds.

Main.Greengrass.update_logger_definition — Method

UpdateLoggerDefinition()

Updates a logger definition.

Required Parameters

LoggerDefinitionId: The ID of the logger definition.

Optional Parameters

Name: The name of the definition.

Main.Greengrass.update_resource_definition — Method

UpdateResourceDefinition()

Updates a resource definition.

Required Parameters

ResourceDefinitionId: The ID of the resource definition.

Optional Parameters

Name: The name of the definition.

Main.Greengrass.update_subscription_definition — Method

UpdateSubscriptionDefinition()

Updates a subscription definition.

Required Parameters

SubscriptionDefinitionId: The ID of the subscription definition.

Optional Parameters

Name: The name of the definition.

Main.Greengrass.update_thing_runtime_configuration — Method

UpdateThingRuntimeConfiguration()

Updates the runtime configuration of a thing.

Required Parameters

ThingName: The thing name.

Optional Parameters

TelemetryConfiguration: Configuration for telemetry service.