Greengrassv2

This page documents function available when using the Greengrassv2 module, created with @service Greengrassv2.

Index

Main.Greengrassv2.batch_associate_client_device_with_core_device

Main.Greengrassv2.batch_disassociate_client_device_from_core_device

Main.Greengrassv2.cancel_deployment

Main.Greengrassv2.create_component_version

Main.Greengrassv2.create_deployment

Main.Greengrassv2.delete_component

Main.Greengrassv2.delete_core_device

Main.Greengrassv2.describe_component

Main.Greengrassv2.get_component

Main.Greengrassv2.get_component_version_artifact

Main.Greengrassv2.get_core_device

Main.Greengrassv2.get_deployment

Main.Greengrassv2.list_client_devices_associated_with_core_device

Main.Greengrassv2.list_component_versions

Main.Greengrassv2.list_components

Main.Greengrassv2.list_core_devices

Main.Greengrassv2.list_deployments

Main.Greengrassv2.list_effective_deployments

Main.Greengrassv2.list_installed_components

Main.Greengrassv2.list_tags_for_resource

Main.Greengrassv2.resolve_component_candidates

Main.Greengrassv2.tag_resource

Main.Greengrassv2.untag_resource

Documentation

Main.Greengrassv2.batch_associate_client_device_with_core_device — Method

batch_associate_client_device_with_core_device(core_device_thing_name)
batch_associate_client_device_with_core_device(core_device_thing_name, params::Dict{String,<:Any})

Associate a list of client devices with a core device. Use this API operation to specify which client devices can discover a core device through cloud discovery. With cloud discovery, client devices connect to IoT Greengrass to retrieve associated core devices' connectivity information and certificates. For more information, see Configure cloud discovery in the IoT Greengrass V2 Developer Guide. Client devices are local IoT devices that connect to and communicate with an IoT Greengrass core device over MQTT. You can connect client devices to a core device to sync MQTT messages and data to Amazon Web Services IoT Core and interact with client devices in Greengrass components. For more information, see Interact with local IoT devices in the IoT Greengrass V2 Developer Guide.

Arguments

core_device_thing_name: The name of the core device. This is also the name of the IoT thing.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"entries": The list of client devices to associate.

Main.Greengrassv2.batch_disassociate_client_device_from_core_device — Method

batch_disassociate_client_device_from_core_device(core_device_thing_name)
batch_disassociate_client_device_from_core_device(core_device_thing_name, params::Dict{String,<:Any})

Disassociate a list of client devices from a core device. After you disassociate a client device from a core device, the client device won't be able to use cloud discovery to retrieve the core device's connectivity information and certificates.

Arguments

core_device_thing_name: The name of the core device. This is also the name of the IoT thing.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"entries": The list of client devices to disassociate.

Main.Greengrassv2.cancel_deployment — Method

cancel_deployment(deployment_id)
cancel_deployment(deployment_id, params::Dict{String,<:Any})

Cancels a deployment. This operation cancels the deployment for devices that haven't yet received it. If a device already received the deployment, this operation doesn't change anything for that device.

Arguments

deployment_id: The ID of the deployment.

Main.Greengrassv2.create_component_version — Method

create_component_version()
create_component_version(params::Dict{String,<:Any})

Creates a component. Components are software that run on Greengrass core devices. After you develop and test a component on your core device, you can use this operation to upload your component to IoT Greengrass. Then, you can deploy the component to other core devices. You can use this operation to do the following: Create components from recipes Create a component from a recipe, which is a file that defines the component's metadata, parameters, dependencies, lifecycle, artifacts, and platform capability. For more information, see IoT Greengrass component recipe reference in the IoT Greengrass V2 Developer Guide. To create a component from a recipe, specify inlineRecipe when you call this operation. Create components from Lambda functions Create a component from an Lambda function that runs on IoT Greengrass. This creates a recipe and artifacts from the Lambda function's deployment package. You can use this operation to migrate Lambda functions from IoT Greengrass V1 to IoT Greengrass V2. This function only accepts Lambda functions that use the following runtimes: Python 2.7 – python2.7 Python 3.7 – python3.7 Python 3.8 – python3.8 Java 8 – java8 Node.js 10 – nodejs10.x Node.js 12 – nodejs12.x To create a component from a Lambda function, specify lambdaFunction when you call this operation.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you can provide to ensure that the request is idempotent. Idempotency means that the request is successfully processed only once, even if you send the request multiple times. When a request succeeds, and you specify the same client token for subsequent successful requests, the IoT Greengrass V2 service returns the successful response that it caches from the previous request. IoT Greengrass V2 caches successful responses for idempotent requests for up to 8 hours.

"inlineRecipe": The recipe to use to create the component. The recipe defines the component's metadata, parameters, dependencies, lifecycle, artifacts, and platform compatibility. You must specify either inlineRecipe or lambdaFunction.

"lambdaFunction": The parameters to create a component from a Lambda function. You must specify either inlineRecipe or lambdaFunction.

"tags": A list of key-value pairs that contain metadata for the resource. For more information, see Tag your resources in the IoT Greengrass V2 Developer Guide.

Main.Greengrassv2.create_deployment — Method

create_deployment(target_arn)
create_deployment(target_arn, params::Dict{String,<:Any})

Creates a continuous deployment for a target, which is a Greengrass core device or group of core devices. When you add a new core device to a group of core devices that has a deployment, IoT Greengrass deploys that group's deployment to the new device. You can define one deployment for each target. When you create a new deployment for a target that has an existing deployment, you replace the previous deployment. IoT Greengrass applies the new deployment to the target devices. Every deployment has a revision number that indicates how many deployment revisions you define for a target. Use this operation to create a new revision of an existing deployment. This operation returns the revision number of the new deployment when you create it. For more information, see the Create deployments in the IoT Greengrass V2 Developer Guide.

Arguments

target_arn: The ARN of the target IoT thing or thing group.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you can provide to ensure that the request is idempotent. Idempotency means that the request is successfully processed only once, even if you send the request multiple times. When a request succeeds, and you specify the same client token for subsequent successful requests, the IoT Greengrass V2 service returns the successful response that it caches from the previous request. IoT Greengrass V2 caches successful responses for idempotent requests for up to 8 hours.

"components": The components to deploy. This is a dictionary, where each key is the name of a component, and each key's value is the version and configuration to deploy for that component.

"deploymentName": The name of the deployment.

"deploymentPolicies": The deployment policies for the deployment. These policies define how the deployment updates components and handles failure.

"iotJobConfiguration": The job configuration for the deployment configuration. The job configuration specifies the rollout, timeout, and stop configurations for the deployment configuration.

"tags": A list of key-value pairs that contain metadata for the resource. For more information, see Tag your resources in the IoT Greengrass V2 Developer Guide.

Main.Greengrassv2.delete_component — Method

delete_component(arn)
delete_component(arn, params::Dict{String,<:Any})

Deletes a version of a component from IoT Greengrass. This operation deletes the component's recipe and artifacts. As a result, deployments that refer to this component version will fail. If you have deployments that use this component version, you can remove the component from the deployment or update the deployment to use a valid version.

Arguments

arn: The ARN of the component version.

Main.Greengrassv2.delete_core_device — Method

delete_core_device(core_device_thing_name)
delete_core_device(core_device_thing_name, params::Dict{String,<:Any})

Deletes a Greengrass core device, which is an IoT thing. This operation removes the core device from the list of core devices. This operation doesn't delete the IoT thing. For more information about how to delete the IoT thing, see DeleteThing in the IoT API Reference.

Arguments

core_device_thing_name: The name of the core device. This is also the name of the IoT thing.

Main.Greengrassv2.describe_component — Method

describe_component(arn)
describe_component(arn, params::Dict{String,<:Any})

Retrieves metadata for a version of a component.

Arguments

arn: The ARN of the component version.

Main.Greengrassv2.get_component — Method

get_component(arn)
get_component(arn, params::Dict{String,<:Any})

Gets the recipe for a version of a component. Core devices can call this operation to identify the artifacts and requirements to install a component.

Arguments

arn: The ARN of the component version.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"recipeOutputFormat": The format of the recipe.

Main.Greengrassv2.get_component_version_artifact — Method

get_component_version_artifact(arn, artifact_name)
get_component_version_artifact(arn, artifact_name, params::Dict{String,<:Any})

Gets the pre-signed URL to download a public component artifact. Core devices call this operation to identify the URL that they can use to download an artifact to install.

Arguments

arn: The ARN of the component version. Specify the ARN of a public component version.

artifact_name: The name of the artifact. You can use the GetComponent operation to download the component recipe, which includes the URI of the artifact. The artifact name is the section of the URI after the scheme. For example, in the artifact URI greengrass:SomeArtifact.zip, the artifact name is SomeArtifact.zip.

Main.Greengrassv2.get_core_device — Method

get_core_device(core_device_thing_name)
get_core_device(core_device_thing_name, params::Dict{String,<:Any})

Retrieves metadata for a Greengrass core device.

Arguments

core_device_thing_name: The name of the core device. This is also the name of the IoT thing.

Main.Greengrassv2.get_deployment — Method

get_deployment(deployment_id)
get_deployment(deployment_id, params::Dict{String,<:Any})

Gets a deployment. Deployments define the components that run on Greengrass core devices.

Arguments

deployment_id: The ID of the deployment.

Main.Greengrassv2.list_client_devices_associated_with_core_device — Method

list_client_devices_associated_with_core_device(core_device_thing_name)
list_client_devices_associated_with_core_device(core_device_thing_name, params::Dict{String,<:Any})

Retrieves a paginated list of client devices that are associated with a core device.

Arguments

core_device_thing_name: The name of the core device. This is also the name of the IoT thing.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to be returned per paginated request.

"nextToken": The token to be used for the next set of paginated results.

Main.Greengrassv2.list_component_versions — Method

list_component_versions(arn)
list_component_versions(arn, params::Dict{String,<:Any})

Retrieves a paginated list of all versions for a component. Greater versions are listed first.

Arguments

arn: The ARN of the component version.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to be returned per paginated request.

"nextToken": The token to be used for the next set of paginated results.

Main.Greengrassv2.list_components — Method

list_components()
list_components(params::Dict{String,<:Any})

Retrieves a paginated list of component summaries. This list includes components that you have permission to view.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to be returned per paginated request.

"nextToken": The token to be used for the next set of paginated results.

"scope": The scope of the components to list. Default: PRIVATE

Main.Greengrassv2.list_core_devices — Method

list_core_devices()
list_core_devices(params::Dict{String,<:Any})

Retrieves a paginated list of Greengrass core devices.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to be returned per paginated request.

"nextToken": The token to be used for the next set of paginated results.

"status": The core device status by which to filter. If you specify this parameter, the list includes only core devices that have this status. Choose one of the following options: HEALTHY – The IoT Greengrass Core software and all components run on the core device without issue. UNHEALTHY – The IoT Greengrass Core software or a component is in a failed state on the core device.

"thingGroupArn": The ARN of the IoT thing group by which to filter. If you specify this parameter, the list includes only core devices that are members of this thing group.

Main.Greengrassv2.list_deployments — Method

list_deployments()
list_deployments(params::Dict{String,<:Any})

Retrieves a paginated list of deployments.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"historyFilter": The filter for the list of deployments. Choose one of the following options: ALL – The list includes all deployments. LATESTONLY – The list includes only the latest revision of each deployment. Default: LATESTONLY

"maxResults": The maximum number of results to be returned per paginated request.

"nextToken": The token to be used for the next set of paginated results.

"targetArn": The ARN of the target IoT thing or thing group.

Main.Greengrassv2.list_effective_deployments — Method

list_effective_deployments(core_device_thing_name)
list_effective_deployments(core_device_thing_name, params::Dict{String,<:Any})

Retrieves a paginated list of deployment jobs that IoT Greengrass sends to Greengrass core devices.

Arguments

core_device_thing_name: The name of the core device. This is also the name of the IoT thing.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to be returned per paginated request.

"nextToken": The token to be used for the next set of paginated results.

Main.Greengrassv2.list_installed_components — Method

list_installed_components(core_device_thing_name)
list_installed_components(core_device_thing_name, params::Dict{String,<:Any})

Retrieves a paginated list of the components that a Greengrass core device runs.

Arguments

core_device_thing_name: The name of the core device. This is also the name of the IoT thing.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to be returned per paginated request.

"nextToken": The token to be used for the next set of paginated results.

Main.Greengrassv2.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Retrieves the list of tags for an IoT Greengrass resource.

Arguments

resource_arn: The ARN of the resource.

Main.Greengrassv2.resolve_component_candidates — Method

resolve_component_candidates(component_candidates, platform)
resolve_component_candidates(component_candidates, platform, params::Dict{String,<:Any})

Retrieves a list of components that meet the component, version, and platform requirements of a deployment. Greengrass core devices call this operation when they receive a deployment to identify the components to install. This operation identifies components that meet all dependency requirements for a deployment. If the requirements conflict, then this operation returns an error and the deployment fails. For example, this occurs if component A requires version &gt;2.0.0 and component B requires version &lt;2.0.0 of a component dependency. When you specify the component candidates to resolve, IoT Greengrass compares each component's digest from the core device with the component's digest in the Amazon Web Services Cloud. If the digests don't match, then IoT Greengrass specifies to use the version from the Amazon Web Services Cloud. To use this operation, you must use the data plane API endpoint and authenticate with an IoT device certificate. For more information, see IoT Greengrass endpoints and quotas.

Arguments

component_candidates: The list of components to resolve.

platform: The platform to use to resolve compatible components.

Main.Greengrassv2.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Adds tags to an IoT Greengrass resource. If a tag already exists for the resource, this operation updates the tag's value.

Arguments

resource_arn: The ARN of the resource to tag.

tags: A list of key-value pairs that contain metadata for the resource. For more information, see Tag your resources in the IoT Greengrass V2 Developer Guide.

Main.Greengrassv2.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes a tag from an IoT Greengrass resource.

Arguments

resource_arn: The ARN of the resource to untag.

tag_keys: A list of keys for tags to remove from the resource.