Groundstation

This page documents function available when using the Groundstation module, created with @service Groundstation.

Index

Main.Groundstation.cancel_contact

Main.Groundstation.create_config

Main.Groundstation.create_dataflow_endpoint_group

Main.Groundstation.create_mission_profile

Main.Groundstation.delete_config

Main.Groundstation.delete_dataflow_endpoint_group

Main.Groundstation.delete_mission_profile

Main.Groundstation.describe_contact

Main.Groundstation.get_config

Main.Groundstation.get_dataflow_endpoint_group

Main.Groundstation.get_minute_usage

Main.Groundstation.get_mission_profile

Main.Groundstation.get_satellite

Main.Groundstation.list_configs

Main.Groundstation.list_contacts

Main.Groundstation.list_dataflow_endpoint_groups

Main.Groundstation.list_ground_stations

Main.Groundstation.list_mission_profiles

Main.Groundstation.list_satellites

Main.Groundstation.list_tags_for_resource

Main.Groundstation.reserve_contact

Main.Groundstation.tag_resource

Main.Groundstation.untag_resource

Main.Groundstation.update_config

Main.Groundstation.update_mission_profile

Documentation

Main.Groundstation.cancel_contact — Method

cancel_contact(contact_id)
cancel_contact(contact_id, params::Dict{String,<:Any})

Cancels a contact with a specified contact ID.

Arguments

contact_id: UUID of a contact.

Main.Groundstation.create_config — Method

create_config(config_data, name)
create_config(config_data, name, params::Dict{String,<:Any})

Creates a Config with the specified configData parameters. Only one type of configData can be specified.

Arguments

config_data: Parameters of a Config.

name: Name of a Config.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"tags": Tags assigned to a Config.

Main.Groundstation.create_dataflow_endpoint_group — Method

create_dataflow_endpoint_group(endpoint_details)
create_dataflow_endpoint_group(endpoint_details, params::Dict{String,<:Any})

Creates a DataflowEndpoint group containing the specified list of DataflowEndpoint objects. The name field in each endpoint is used in your mission profile DataflowEndpointConfig to specify which endpoints to use during a contact. When a contact uses multiple DataflowEndpointConfig objects, each Config must match a DataflowEndpoint in the same group.

Arguments

endpoint_details: Endpoint details of each endpoint in the dataflow endpoint group.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"tags": Tags of a dataflow endpoint group.

Main.Groundstation.create_mission_profile — Method

create_mission_profile(dataflow_edges, minimum_viable_contact_duration_seconds, name, tracking_config_arn)
create_mission_profile(dataflow_edges, minimum_viable_contact_duration_seconds, name, tracking_config_arn, params::Dict{String,<:Any})

Creates a mission profile. dataflowEdges is a list of lists of strings. Each lower level list of strings has two elements: a from ARN and a to ARN.

Arguments

dataflow_edges: A list of lists of ARNs. Each list of ARNs is an edge, with a from Config and a to Config.

minimum_viable_contact_duration_seconds: Smallest amount of time in seconds that you’d like to see for an available contact. AWS Ground Station will not present you with contacts shorter than this duration.

name: Name of a mission profile.

tracking_config_arn: ARN of a tracking Config.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"contactPostPassDurationSeconds": Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

"contactPrePassDurationSeconds": Amount of time prior to contact start you’d like to receive a CloudWatch event indicating an upcoming pass.

"tags": Tags assigned to a mission profile.

Main.Groundstation.delete_config — Method

delete_config(config_id, config_type)
delete_config(config_id, config_type, params::Dict{String,<:Any})

Deletes a Config.

Arguments

config_id: UUID of a Config.

config_type: Type of a Config.

Main.Groundstation.delete_dataflow_endpoint_group — Method

delete_dataflow_endpoint_group(dataflow_endpoint_group_id)
delete_dataflow_endpoint_group(dataflow_endpoint_group_id, params::Dict{String,<:Any})

Deletes a dataflow endpoint group.

Arguments

dataflow_endpoint_group_id: UUID of a dataflow endpoint group.

Main.Groundstation.delete_mission_profile — Method

delete_mission_profile(mission_profile_id)
delete_mission_profile(mission_profile_id, params::Dict{String,<:Any})

Deletes a mission profile.

Arguments

mission_profile_id: UUID of a mission profile.

Main.Groundstation.describe_contact — Method

describe_contact(contact_id)
describe_contact(contact_id, params::Dict{String,<:Any})

Describes an existing contact.

Arguments

contact_id: UUID of a contact.

Main.Groundstation.get_config — Method

get_config(config_id, config_type)
get_config(config_id, config_type, params::Dict{String,<:Any})

Returns Config information. Only one Config response can be returned.

Arguments

config_id: UUID of a Config.

config_type: Type of a Config.

Main.Groundstation.get_dataflow_endpoint_group — Method

get_dataflow_endpoint_group(dataflow_endpoint_group_id)
get_dataflow_endpoint_group(dataflow_endpoint_group_id, params::Dict{String,<:Any})

Returns the dataflow endpoint group.

Arguments

dataflow_endpoint_group_id: UUID of a dataflow endpoint group.

Main.Groundstation.get_minute_usage — Method

get_minute_usage(month, year)
get_minute_usage(month, year, params::Dict{String,<:Any})

Returns the number of minutes used by account.

Arguments

month: The month being requested, with a value of 1-12.

year: The year being requested, in the format of YYYY.

Main.Groundstation.get_mission_profile — Method

get_mission_profile(mission_profile_id)
get_mission_profile(mission_profile_id, params::Dict{String,<:Any})

Returns a mission profile.

Arguments

mission_profile_id: UUID of a mission profile.

Main.Groundstation.get_satellite — Method

get_satellite(satellite_id)
get_satellite(satellite_id, params::Dict{String,<:Any})

Returns a satellite.

Arguments

satellite_id: UUID of a satellite.

Main.Groundstation.list_configs — Method

list_configs()
list_configs(params::Dict{String,<:Any})

Returns a list of Config objects.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Maximum number of Configs returned.

"nextToken": Next token returned in the request of a previous ListConfigs call. Used to get the next page of results.

Main.Groundstation.list_contacts — Method

list_contacts(end_time, start_time, status_list)
list_contacts(end_time, start_time, status_list, params::Dict{String,<:Any})

Returns a list of contacts. If statusList contains AVAILABLE, the request must include groundStation, missionprofileArn, and satelliteArn.

Arguments

end_time: End time of a contact.

start_time: Start time of a contact.

status_list: Status of a contact reservation.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"groundStation": Name of a ground station.

"maxResults": Maximum number of contacts returned.

"missionProfileArn": ARN of a mission profile.

"nextToken": Next token returned in the request of a previous ListContacts call. Used to get the next page of results.

"satelliteArn": ARN of a satellite.

Main.Groundstation.list_dataflow_endpoint_groups — Method

list_dataflow_endpoint_groups()
list_dataflow_endpoint_groups(params::Dict{String,<:Any})

Returns a list of DataflowEndpoint groups.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Maximum number of dataflow endpoint groups returned.

"nextToken": Next token returned in the request of a previous ListDataflowEndpointGroups call. Used to get the next page of results.

Main.Groundstation.list_ground_stations — Method

list_ground_stations()
list_ground_stations(params::Dict{String,<:Any})

Returns a list of ground stations.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Maximum number of ground stations returned.

"nextToken": Next token that can be supplied in the next call to get the next page of ground stations.

"satelliteId": Satellite ID to retrieve on-boarded ground stations.

Main.Groundstation.list_mission_profiles — Method

list_mission_profiles()
list_mission_profiles(params::Dict{String,<:Any})

Returns a list of mission profiles.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Maximum number of mission profiles returned.

"nextToken": Next token returned in the request of a previous ListMissionProfiles call. Used to get the next page of results.

Main.Groundstation.list_satellites — Method

list_satellites()
list_satellites(params::Dict{String,<:Any})

Returns a list of satellites.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Maximum number of satellites returned.

"nextToken": Next token that can be supplied in the next call to get the next page of satellites.

Main.Groundstation.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Returns a list of tags for a specified resource.

Arguments

resource_arn: ARN of a resource.

Main.Groundstation.reserve_contact — Method

reserve_contact(end_time, ground_station, mission_profile_arn, satellite_arn, start_time)
reserve_contact(end_time, ground_station, mission_profile_arn, satellite_arn, start_time, params::Dict{String,<:Any})

Reserves a contact using specified parameters.

Arguments

end_time: End time of a contact.

ground_station: Name of a ground station.

mission_profile_arn: ARN of a mission profile.

satellite_arn: ARN of a satellite

start_time: Start time of a contact.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"tags": Tags assigned to a contact.

Main.Groundstation.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Assigns a tag to a resource.

Arguments

resource_arn: ARN of a resource tag.

tags: Tags assigned to a resource.

Main.Groundstation.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Deassigns a resource tag.

Arguments

resource_arn: ARN of a resource.

tag_keys: Keys of a resource tag.

Main.Groundstation.update_config — Method

update_config(config_data, config_id, config_type, name)
update_config(config_data, config_id, config_type, name, params::Dict{String,<:Any})

Updates the Config used when scheduling contacts. Updating a Config will not update the execution parameters for existing future contacts scheduled with this Config.

Arguments

config_data: Parameters of a Config.

config_id: UUID of a Config.

config_type: Type of a Config.

name: Name of a Config.

Main.Groundstation.update_mission_profile — Method

update_mission_profile(mission_profile_id)
update_mission_profile(mission_profile_id, params::Dict{String,<:Any})

Updates a mission profile. Updating a mission profile will not update the execution parameters for existing future contacts.

Arguments

mission_profile_id: UUID of a mission profile.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"contactPostPassDurationSeconds": Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

"contactPrePassDurationSeconds": Amount of time after a contact ends that you’d like to receive a CloudWatch event indicating the pass has finished.

"dataflowEdges": A list of lists of ARNs. Each list of ARNs is an edge, with a from Config and a to Config.

"minimumViableContactDurationSeconds": Smallest amount of time in seconds that you’d like to see for an available contact. AWS Ground Station will not present you with contacts shorter than this duration.

"name": Name of a mission profile.

"trackingConfigArn": ARN of a tracking Config.