Iot Data Plane

This page documents function available when using the Iot_Data_Plane module, created with @service Iot_Data_Plane.

Index

Main.Iot_Data_Plane.delete_thing_shadow

Main.Iot_Data_Plane.get_retained_message

Main.Iot_Data_Plane.get_thing_shadow

Main.Iot_Data_Plane.list_named_shadows_for_thing

Main.Iot_Data_Plane.list_retained_messages

Main.Iot_Data_Plane.publish

Main.Iot_Data_Plane.update_thing_shadow

Documentation

Main.Iot_Data_Plane.delete_thing_shadow — Method

delete_thing_shadow(thing_name)
delete_thing_shadow(thing_name, params::Dict{String,<:Any})

Deletes the shadow for the specified thing. Requires permission to access the DeleteThingShadow action. For more information, see DeleteThingShadow in the IoT Developer Guide.

Arguments

thing_name: The name of the thing.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"name": The name of the shadow.

Main.Iot_Data_Plane.get_retained_message — Method

get_retained_message(topic)
get_retained_message(topic, params::Dict{String,<:Any})

Gets the details of a single retained message for the specified topic. This action returns the message payload of the retained message, which can incur messaging costs. To list only the topic names of the retained messages, call ListRetainedMessages. Requires permission to access the GetRetainedMessage action. For more information about messaging costs, see IoT Core pricing - Messaging.

Arguments

topic: The topic name of the retained message to retrieve.

Main.Iot_Data_Plane.get_thing_shadow — Method

get_thing_shadow(thing_name)
get_thing_shadow(thing_name, params::Dict{String,<:Any})

Gets the shadow for the specified thing. Requires permission to access the GetThingShadow action. For more information, see GetThingShadow in the IoT Developer Guide.

Arguments

thing_name: The name of the thing.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"name": The name of the shadow.

Main.Iot_Data_Plane.list_named_shadows_for_thing — Method

list_named_shadows_for_thing(thing_name)
list_named_shadows_for_thing(thing_name, params::Dict{String,<:Any})

Lists the shadows for the specified thing. Requires permission to access the ListNamedShadowsForThing action.

Arguments

thing_name: The name of the thing.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nextToken": The token to retrieve the next set of results.

"pageSize": The result page size.

Main.Iot_Data_Plane.list_retained_messages — Method

list_retained_messages()
list_retained_messages(params::Dict{String,<:Any})

Lists summary information about the retained messages stored for the account. This action returns only the topic names of the retained messages. It doesn't return any message payloads. Although this action doesn't return a message payload, it can still incur messaging costs. To get the message payload of a retained message, call GetRetainedMessage with the topic name of the retained message. Requires permission to access the ListRetainedMessages action. For more information about messaging costs, see IoT Core pricing - Messaging.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return at one time.

"nextToken": To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

Main.Iot_Data_Plane.publish — Method

publish(topic)
publish(topic, params::Dict{String,<:Any})

Publishes an MQTT message. Requires permission to access the Publish action. For more information about MQTT messages, see MQTT Protocol in the IoT Developer Guide. For more information about messaging costs, see IoT Core pricing - Messaging.

Arguments

topic: The name of the MQTT topic.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"payload": The message body. MQTT accepts text, binary, and empty (null) message payloads. Publishing an empty (null) payload with retain = true deletes the retained message identified by topic from IoT Core.

"qos": The Quality of Service (QoS) level.

"retain": A Boolean value that determines whether to set the RETAIN flag when the message is published. Setting the RETAIN flag causes the message to be retained and sent to new subscribers to the topic. Valid values: true | false Default value: false

Main.Iot_Data_Plane.update_thing_shadow — Method

update_thing_shadow(payload, thing_name)
update_thing_shadow(payload, thing_name, params::Dict{String,<:Any})

Updates the shadow for the specified thing. Requires permission to access the UpdateThingShadow action. For more information, see UpdateThingShadow in the IoT Developer Guide.

Arguments

payload: The state information, in JSON format.

thing_name: The name of the thing.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"name": The name of the shadow.