Iot Events Data

This page documents function available when using the Iot_Events_Data module, created with @service Iot_Events_Data.

Index

Main.Iot_Events_Data.batch_acknowledge_alarm

Main.Iot_Events_Data.batch_delete_detector

Main.Iot_Events_Data.batch_disable_alarm

Main.Iot_Events_Data.batch_enable_alarm

Main.Iot_Events_Data.batch_put_message

Main.Iot_Events_Data.batch_reset_alarm

Main.Iot_Events_Data.batch_snooze_alarm

Main.Iot_Events_Data.batch_update_detector

Main.Iot_Events_Data.describe_alarm

Main.Iot_Events_Data.describe_detector

Main.Iot_Events_Data.list_alarms

Main.Iot_Events_Data.list_detectors

Documentation

Main.Iot_Events_Data.batch_acknowledge_alarm — Method

batch_acknowledge_alarm(acknowledge_action_requests)
batch_acknowledge_alarm(acknowledge_action_requests, params::Dict{String,<:Any})

Acknowledges one or more alarms. The alarms change to the ACKNOWLEDGED state after you acknowledge them.

Arguments

acknowledge_action_requests: The list of acknowledge action requests. You can specify up to 10 requests per operation.

Main.Iot_Events_Data.batch_delete_detector — Method

batch_delete_detector(detectors)
batch_delete_detector(detectors, params::Dict{String,<:Any})

Deletes one or more detectors that were created. When a detector is deleted, its state will be cleared and the detector will be removed from the list of detectors. The deleted detector will no longer appear if referenced in the ListDetectors API call.

Arguments

detectors: The list of one or more detectors to be deleted.

Main.Iot_Events_Data.batch_disable_alarm — Method

batch_disable_alarm(disable_action_requests)
batch_disable_alarm(disable_action_requests, params::Dict{String,<:Any})

Disables one or more alarms. The alarms change to the DISABLED state after you disable them.

Arguments

disable_action_requests: The list of disable action requests. You can specify up to 10 requests per operation.

Main.Iot_Events_Data.batch_enable_alarm — Method

batch_enable_alarm(enable_action_requests)
batch_enable_alarm(enable_action_requests, params::Dict{String,<:Any})

Enables one or more alarms. The alarms change to the NORMAL state after you enable them.

Arguments

enable_action_requests: The list of enable action requests. You can specify up to 10 requests per operation.

Main.Iot_Events_Data.batch_put_message — Method

batch_put_message(messages)
batch_put_message(messages, params::Dict{String,<:Any})

Sends a set of messages to the IoT Events system. Each message payload is transformed into the input you specify ("inputName") and ingested into any detectors that monitor that input. If multiple messages are sent, the order in which the messages are processed isn't guaranteed. To guarantee ordering, you must send messages one at a time and wait for a successful response.

Arguments

messages: The list of messages to send. Each message has the following format: '{ "messageId": "string", "inputName": "string", "payload": "string"}'

Main.Iot_Events_Data.batch_reset_alarm — Method

batch_reset_alarm(reset_action_requests)
batch_reset_alarm(reset_action_requests, params::Dict{String,<:Any})

Resets one or more alarms. The alarms return to the NORMAL state after you reset them.

Arguments

reset_action_requests: The list of reset action requests. You can specify up to 10 requests per operation.

Main.Iot_Events_Data.batch_snooze_alarm — Method

batch_snooze_alarm(snooze_action_requests)
batch_snooze_alarm(snooze_action_requests, params::Dict{String,<:Any})

Changes one or more alarms to the snooze mode. The alarms change to the SNOOZE_DISABLED state after you set them to the snooze mode.

Arguments

snooze_action_requests: The list of snooze action requests. You can specify up to 10 requests per operation.

Main.Iot_Events_Data.batch_update_detector — Method

batch_update_detector(detectors)
batch_update_detector(detectors, params::Dict{String,<:Any})

Updates the state, variable values, and timer settings of one or more detectors (instances) of a specified detector model.

Arguments

detectors: The list of detectors (instances) to update, along with the values to update.

Main.Iot_Events_Data.describe_alarm — Method

describe_alarm(alarm_model_name)
describe_alarm(alarm_model_name, params::Dict{String,<:Any})

Retrieves information about an alarm.

Arguments

alarm_model_name: The name of the alarm model.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"keyValue": The value of the key used as a filter to select only the alarms associated with the key.

Main.Iot_Events_Data.describe_detector — Method

describe_detector(detector_model_name)
describe_detector(detector_model_name, params::Dict{String,<:Any})

Returns information about the specified detector (instance).

Arguments

detector_model_name: The name of the detector model whose detectors (instances) you want information about.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"keyValue": A filter used to limit results to detectors (instances) created because of the given key ID.

Main.Iot_Events_Data.list_alarms — Method

list_alarms(alarm_model_name)
list_alarms(alarm_model_name, params::Dict{String,<:Any})

Lists one or more alarms. The operation returns only the metadata associated with each alarm.

Arguments

alarm_model_name: The name of the alarm model.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to be returned per request.

"nextToken": The token that you can use to return the next set of results.

Main.Iot_Events_Data.list_detectors — Method

list_detectors(detector_model_name)
list_detectors(detector_model_name, params::Dict{String,<:Any})

Lists detectors (the instances of a detector model).

Arguments

detector_model_name: The name of the detector model whose detectors (instances) are listed.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to be returned per request.

"nextToken": The token that you can use to return the next set of results.

"stateName": A filter that limits results to those detectors (instances) in the given state.