Iot Managed Integrations

This page documents function available when using the Iot_Managed_Integrations module, created with @service Iot_Managed_Integrations.

Index

Main.Iot_Managed_Integrations.create_account_association

Main.Iot_Managed_Integrations.create_cloud_connector

Main.Iot_Managed_Integrations.create_connector_destination

Main.Iot_Managed_Integrations.create_credential_locker

Main.Iot_Managed_Integrations.create_destination

Main.Iot_Managed_Integrations.create_event_log_configuration

Main.Iot_Managed_Integrations.create_managed_thing

Main.Iot_Managed_Integrations.create_notification_configuration

Main.Iot_Managed_Integrations.create_ota_task

Main.Iot_Managed_Integrations.create_ota_task_configuration

Main.Iot_Managed_Integrations.create_provisioning_profile

Main.Iot_Managed_Integrations.delete_account_association

Main.Iot_Managed_Integrations.delete_cloud_connector

Main.Iot_Managed_Integrations.delete_connector_destination

Main.Iot_Managed_Integrations.delete_credential_locker

Main.Iot_Managed_Integrations.delete_destination

Main.Iot_Managed_Integrations.delete_event_log_configuration

Main.Iot_Managed_Integrations.delete_managed_thing

Main.Iot_Managed_Integrations.delete_notification_configuration

Main.Iot_Managed_Integrations.delete_ota_task

Main.Iot_Managed_Integrations.delete_ota_task_configuration

Main.Iot_Managed_Integrations.delete_provisioning_profile

Main.Iot_Managed_Integrations.deregister_account_association

Main.Iot_Managed_Integrations.get_account_association

Main.Iot_Managed_Integrations.get_cloud_connector

Main.Iot_Managed_Integrations.get_connector_destination

Main.Iot_Managed_Integrations.get_credential_locker

Main.Iot_Managed_Integrations.get_custom_endpoint

Main.Iot_Managed_Integrations.get_default_encryption_configuration

Main.Iot_Managed_Integrations.get_destination

Main.Iot_Managed_Integrations.get_device_discovery

Main.Iot_Managed_Integrations.get_event_log_configuration

Main.Iot_Managed_Integrations.get_hub_configuration

Main.Iot_Managed_Integrations.get_managed_thing

Main.Iot_Managed_Integrations.get_managed_thing_capabilities

Main.Iot_Managed_Integrations.get_managed_thing_certificate

Main.Iot_Managed_Integrations.get_managed_thing_connectivity_data

Main.Iot_Managed_Integrations.get_managed_thing_meta_data

Main.Iot_Managed_Integrations.get_managed_thing_state

Main.Iot_Managed_Integrations.get_notification_configuration

Main.Iot_Managed_Integrations.get_ota_task

Main.Iot_Managed_Integrations.get_ota_task_configuration

Main.Iot_Managed_Integrations.get_provisioning_profile

Main.Iot_Managed_Integrations.get_runtime_log_configuration

Main.Iot_Managed_Integrations.get_schema_version

Main.Iot_Managed_Integrations.list_account_associations

Main.Iot_Managed_Integrations.list_cloud_connectors

Main.Iot_Managed_Integrations.list_connector_destinations

Main.Iot_Managed_Integrations.list_credential_lockers

Main.Iot_Managed_Integrations.list_destinations

Main.Iot_Managed_Integrations.list_device_discoveries

Main.Iot_Managed_Integrations.list_discovered_devices

Main.Iot_Managed_Integrations.list_event_log_configurations

Main.Iot_Managed_Integrations.list_managed_thing_account_associations

Main.Iot_Managed_Integrations.list_managed_thing_schemas

Main.Iot_Managed_Integrations.list_managed_things

Main.Iot_Managed_Integrations.list_notification_configurations

Main.Iot_Managed_Integrations.list_ota_task_configurations

Main.Iot_Managed_Integrations.list_ota_task_executions

Main.Iot_Managed_Integrations.list_ota_tasks

Main.Iot_Managed_Integrations.list_provisioning_profiles

Main.Iot_Managed_Integrations.list_schema_versions

Main.Iot_Managed_Integrations.list_tags_for_resource

Main.Iot_Managed_Integrations.put_default_encryption_configuration

Main.Iot_Managed_Integrations.put_hub_configuration

Main.Iot_Managed_Integrations.put_runtime_log_configuration

Main.Iot_Managed_Integrations.register_account_association

Main.Iot_Managed_Integrations.register_custom_endpoint

Main.Iot_Managed_Integrations.reset_runtime_log_configuration

Main.Iot_Managed_Integrations.send_connector_event

Main.Iot_Managed_Integrations.send_managed_thing_command

Main.Iot_Managed_Integrations.start_account_association_refresh

Main.Iot_Managed_Integrations.start_device_discovery

Main.Iot_Managed_Integrations.tag_resource

Main.Iot_Managed_Integrations.untag_resource

Main.Iot_Managed_Integrations.update_account_association

Main.Iot_Managed_Integrations.update_cloud_connector

Main.Iot_Managed_Integrations.update_connector_destination

Main.Iot_Managed_Integrations.update_destination

Main.Iot_Managed_Integrations.update_event_log_configuration

Main.Iot_Managed_Integrations.update_managed_thing

Main.Iot_Managed_Integrations.update_notification_configuration

Main.Iot_Managed_Integrations.update_ota_task

Documentation

Main.Iot_Managed_Integrations.create_account_association — Function

create_account_association(connector_destination_id)
create_account_association(connector_destination_id, params::Dict{String,<:Any})

Creates a new account association via the destination id.

Arguments

connector_destination_id: The identifier of the connector destination.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": An idempotency token. If you retry a request that completed successfully initially using the same client token and parameters, then the retry attempt will succeed without performing any further actions.

"Description": A description of the account association request.

"GeneralAuthorization": The General Authorization reference by authorization material name.

"Name": The name of the destination for the new account association.

"Tags": A set of key/value pairs that are used to manage the account association.

Main.Iot_Managed_Integrations.create_cloud_connector — Function

create_cloud_connector(endpoint_config, name)
create_cloud_connector(endpoint_config, name, params::Dict{String,<:Any})

Creates a C2C (cloud-to-cloud) connector.

Arguments

endpoint_config: The configuration details for the cloud connector endpoint, including connection parameters and authentication requirements.

name: The display name of the C2C connector.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": An idempotency token. If you retry a request that completed successfully initially using the same client token and parameters, then the retry attempt will succeed without performing any further actions.

"Description": A description of the C2C connector.

"EndpointType": The type of endpoint used for the cloud connector, which defines how the connector communicates with external services.

Main.Iot_Managed_Integrations.create_connector_destination — Function

create_connector_destination(auth_config, cloud_connector_id)
create_connector_destination(auth_config, cloud_connector_id, params::Dict{String,<:Any})

Create a connector destination for connecting a cloud-to-cloud (C2C) connector to the customer's Amazon Web Services account.

Arguments

auth_config: The authentication configuration details for the connector destination, including OAuth settings and other authentication parameters.

cloud_connector_id: The identifier of the C2C connector.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AuthType": The authentication type used for the connector destination, which determines how credentials and access are managed.

"ClientToken": An idempotency token. If you retry a request that completed successfully initially using the same client token and parameters, then the retry attempt will succeed without performing any further actions.

"Description": A description of the connector destination.

"Name": The display name of the connector destination.

"SecretsManager": The AWS Secrets Manager configuration used to securely store and manage sensitive information for the connector destination.

Main.Iot_Managed_Integrations.create_credential_locker — Function

create_credential_locker()
create_credential_locker(params::Dict{String,<:Any})

Create a credential locker.

Note

This operation will not trigger the creation of all the manufacturing resources.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": An idempotency token. If you retry a request that completed successfully initially using the same client token and parameters, then the retry attempt will succeed without performing any further actions.

"Name": The name of the credential locker.

"Tags": A set of key/value pairs that are used to manage the credential locker.

Main.Iot_Managed_Integrations.create_destination — Function

create_destination(delivery_destination_arn, delivery_destination_type, name, role_arn)
create_destination(delivery_destination_arn, delivery_destination_type, name, role_arn, params::Dict{String,<:Any})

Create a notification destination such as Kinesis Data Streams that receive events and notifications from Managed integrations. Managed integrations uses the destination to determine where to deliver notifications.

Arguments

delivery_destination_arn: The Amazon Resource Name (ARN) of the customer-managed destination.

delivery_destination_type: The destination type for the customer-managed destination.

name: The name of the customer-managed destination.

role_arn: The Amazon Resource Name (ARN) of the delivery destination role.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": An idempotency token. If you retry a request that completed successfully initially using the same client token and parameters, then the retry attempt will succeed without performing any further actions.

"Description": The description of the customer-managed destination.

"Tags": A set of key/value pairs that are used to manage the destination.

Main.Iot_Managed_Integrations.create_event_log_configuration — Function

create_event_log_configuration(event_log_level, resource_type)
create_event_log_configuration(event_log_level, resource_type, params::Dict{String,<:Any})

Set the event log configuration for the account, resource type, or specific resource.

Arguments

event_log_level: The logging level for the event log configuration.

resource_type: The type of resource for the event log configuration.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": An idempotency token. If you retry a request that completed successfully initially using the same client token and parameters, then the retry attempt will succeed without performing any further actions.

"ResourceId": The identifier of the resource for the event log configuration.

Main.Iot_Managed_Integrations.create_managed_thing — Function

create_managed_thing(authentication_material, authentication_material_type, role)
create_managed_thing(authentication_material, authentication_material_type, role, params::Dict{String,<:Any})

Creates a managed thing. A managed thing contains the device identifier, protocol supported, and capabilities of the device in a data model format defined by Managed integrations.

Arguments

authentication_material: The authentication material defining the device connectivity setup requests. The authorization materials used are the device bar code.

authentication_material_type: The type of authentication material used for device connectivity setup requests.

role: The type of device used. This will be the hub controller, cloud device, or AWS IoT device.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Brand": The brand of the device.

"Capabilities": The capabilities of the device such as light bulb.

"CapabilityReport": A report of the capabilities for the managed thing.

"CapabilitySchemas": The capability schemas that define the functionality and features supported by the managed thing, including device capabilities and their associated properties.

"Classification": The classification of the managed thing such as light bulb or thermostat.

"ClientToken": An idempotency token. If you retry a request that completed successfully initially using the same client token and parameters, then the retry attempt will succeed without performing any further actions.

"CredentialLockerId": The identifier of the credential for the managed thing.

"MetaData": The metadata for the managed thing.

Note

The managedThing metadata parameter is used for associating attributes with a managedThing that can be used for grouping over-the-air (OTA) tasks. Name value pairs in metadata can be used in the OtaTargetQueryString parameter for the CreateOtaTask API operation.

"Model": The model of the device.

"Name": The name of the managed thing representing the physical device.

"Owner": Owner of the device, usually an indication of whom the device belongs to. This value should not contain personal identifiable information.

"SerialNumber": The serial number of the device.

"Tags": A set of key/value pairs that are used to manage the managed thing.

"WiFiSimpleSetupConfiguration": The Wi-Fi Simple Setup configuration for the managed thing, which defines provisioning capabilities and timeout settings.

Main.Iot_Managed_Integrations.create_notification_configuration — Function

create_notification_configuration(destination_name, event_type)
create_notification_configuration(destination_name, event_type, params::Dict{String,<:Any})

Creates a notification configuration. A configuration is a connection between an event type and a destination that you have already created.

Arguments

destination_name: The name of the destination for the notification configuration.

event_type: The type of event triggering a device notification to the customer-managed destination.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": An idempotency token. If you retry a request that completed successfully initially using the same client token and parameters, then the retry attempt will succeed without performing any further actions.

"Tags": A set of key/value pairs that are used to manage the notification configuration.

Main.Iot_Managed_Integrations.create_ota_task — Function

create_ota_task(ota_type, s3_url)
create_ota_task(ota_type, s3_url, params::Dict{String,<:Any})

Create an over-the-air (OTA) task to target a device.

Arguments

ota_type: The frequency type for the over-the-air (OTA) task.

s3_url: The URL to the Amazon S3 bucket where the over-the-air (OTA) task is stored.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": An idempotency token. If you retry a request that completed successfully initially using the same client token and parameters, then the retry attempt will succeed without performing any further actions.

"Description": The description of the over-the-air (OTA) task.

"OtaMechanism": The deployment mechanism for the over-the-air (OTA) task.

"OtaSchedulingConfig":

"OtaTargetQueryString": The query string to add things to the thing group.

"OtaTaskExecutionRetryConfig":

"Protocol": The connection protocol the over-the-air (OTA) task uses to update the device.

"Tags": A set of key/value pairs that are used to manage the over-the-air (OTA) task.

"Target": The device targeted for the over-the-air (OTA) task.

"TaskConfigurationId": The identifier for the over-the-air (OTA) task configuration.

Main.Iot_Managed_Integrations.create_ota_task_configuration — Function

create_ota_task_configuration()
create_ota_task_configuration(params::Dict{String,<:Any})

Create a configuraiton for the over-the-air (OTA) task.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": An idempotency token. If you retry a request that completed successfully initially using the same client token and parameters, then the retry attempt will succeed without performing any further actions.

"Description": A description of the over-the-air (OTA) task configuration.

"Name": The name of the over-the-air (OTA) task.

"PushConfig": Describes the type of configuration used for the over-the-air (OTA) task.

Main.Iot_Managed_Integrations.create_provisioning_profile — Function

create_provisioning_profile(provisioning_type)
create_provisioning_profile(provisioning_type, params::Dict{String,<:Any})

Create a provisioning profile for a device to execute the provisioning flows using a provisioning template. The provisioning template is a document that defines the set of resources and policies applied to a device during the provisioning process.

Arguments

provisioning_type: The type of provisioning workflow the device uses for onboarding to IoT managed integrations.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"CaCertificate": The id of the certificate authority (CA) certificate.

"ClaimCertificate": The claim certificate.

"ClientToken": An idempotency token. If you retry a request that completed successfully initially using the same client token and parameters, then the retry attempt will succeed without performing any further actions.

"Name": The name of the provisioning template.

"Tags": A set of key/value pairs that are used to manage the provisioning profile.

Main.Iot_Managed_Integrations.delete_account_association — Function

delete_account_association(account_association_id)
delete_account_association(account_association_id, params::Dict{String,<:Any})

Remove a third-party account association for an end user.

Note

You must first call the DeregisterAccountAssociation to remove the connection between the managed thing and the third-party account before calling the DeleteAccountAssociation API.

Arguments

account_association_id: The unique identifier of the account association to be deleted.

Main.Iot_Managed_Integrations.delete_cloud_connector — Function

delete_cloud_connector(identifier)
delete_cloud_connector(identifier, params::Dict{String,<:Any})

Delete a cloud connector.

Arguments

identifier: The identifier of the cloud connector.

Main.Iot_Managed_Integrations.delete_connector_destination — Function

delete_connector_destination(identifier)
delete_connector_destination(identifier, params::Dict{String,<:Any})

Delete a connector destination linked to a cloud-to-cloud (C2C) connector.

Note

Deletion can't be done if the account association has used this connector destination.

Arguments

identifier: The identifier of the connector destination.

Main.Iot_Managed_Integrations.delete_credential_locker — Function

delete_credential_locker(identifier)
delete_credential_locker(identifier, params::Dict{String,<:Any})

Delete a credential locker.

Note

This operation can't be undone and any existing device won't be able to use IoT managed integrations.

Arguments

identifier: The identifier of the credential locker.

Main.Iot_Managed_Integrations.delete_destination — Function

delete_destination(name)
delete_destination(name, params::Dict{String,<:Any})

Deletes a notification destination specified by name.

Arguments

name: The id of the customer-managed destination.

Main.Iot_Managed_Integrations.delete_event_log_configuration — Function

delete_event_log_configuration(id)
delete_event_log_configuration(id, params::Dict{String,<:Any})

Delete an event log configuration.

Arguments

id: The identifier of the event log configuration.

Main.Iot_Managed_Integrations.delete_managed_thing — Function

delete_managed_thing(identifier)
delete_managed_thing(identifier, params::Dict{String,<:Any})

Delete a managed thing. For direct-connected and hub-connected devices connecting with Managed integrations via a controller, all of the devices connected to it will have their status changed to PENDING. It is not possible to remove a cloud-to-cloud device.

Arguments

identifier: The id of the managed thing.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Force": When set to TRUE, a forceful deteletion of the managed thing will occur. When set to FALSE, a non-forceful deletion of the managed thing will occur.

Main.Iot_Managed_Integrations.delete_notification_configuration — Function

delete_notification_configuration(event_type)
delete_notification_configuration(event_type, params::Dict{String,<:Any})

Deletes a notification configuration.

Arguments

event_type: The type of event triggering a device notification to the customer-managed destination.

Main.Iot_Managed_Integrations.delete_ota_task — Function

delete_ota_task(identifier)
delete_ota_task(identifier, params::Dict{String,<:Any})

Delete the over-the-air (OTA) task.

Arguments

identifier: The identifier of the over-the-air (OTA) task.

Main.Iot_Managed_Integrations.delete_ota_task_configuration — Function

delete_ota_task_configuration(identifier)
delete_ota_task_configuration(identifier, params::Dict{String,<:Any})

Delete the over-the-air (OTA) task configuration.

Arguments

identifier: The identifier of the over-the-air (OTA) task configuration.

Main.Iot_Managed_Integrations.delete_provisioning_profile — Function

delete_provisioning_profile(identifier)
delete_provisioning_profile(identifier, params::Dict{String,<:Any})

Delete a provisioning profile.

Arguments

identifier: The name of the provisioning template.

Main.Iot_Managed_Integrations.deregister_account_association — Function

deregister_account_association(account_association_id, managed_thing_id)
deregister_account_association(account_association_id, managed_thing_id, params::Dict{String,<:Any})

Deregister an account association from a managed thing.

Arguments

account_association_id: The unique identifier of the account association to be deregistered.

managed_thing_id: The identifier of the managed thing to be deregistered from the account association.

Main.Iot_Managed_Integrations.get_account_association — Function

get_account_association(account_association_id)
get_account_association(account_association_id, params::Dict{String,<:Any})

Get an account association for an Amazon Web Services account linked to a customer-managed destination.

Arguments

account_association_id: The unique identifier of the account association to retrieve.

Main.Iot_Managed_Integrations.get_cloud_connector — Function

get_cloud_connector(identifier)
get_cloud_connector(identifier, params::Dict{String,<:Any})

Get configuration details for a cloud connector.

Arguments

identifier: The identifier of the C2C connector.

Main.Iot_Managed_Integrations.get_connector_destination — Function

get_connector_destination(identifier)
get_connector_destination(identifier, params::Dict{String,<:Any})

Get connector destination details linked to a cloud-to-cloud (C2C) connector.

Arguments

identifier: The identifier of the C2C connector destination.

Main.Iot_Managed_Integrations.get_credential_locker — Function

get_credential_locker(identifier)
get_credential_locker(identifier, params::Dict{String,<:Any})

Get information on an existing credential locker

Arguments

identifier: The identifier of the credential locker.

Main.Iot_Managed_Integrations.get_custom_endpoint — Function

get_custom_endpoint()
get_custom_endpoint(params::Dict{String,<:Any})

Returns the IoT managed integrations custom endpoint.

Main.Iot_Managed_Integrations.get_default_encryption_configuration — Function

get_default_encryption_configuration()
get_default_encryption_configuration(params::Dict{String,<:Any})

Retrieves information about the default encryption configuration for the Amazon Web Services account in the default or specified region. For more information, see Key management in the AWS IoT SiteWise User Guide.

Main.Iot_Managed_Integrations.get_destination — Function

get_destination(name)
get_destination(name, params::Dict{String,<:Any})

Gets a destination by name.

Arguments

name: The name of the customer-managed destination.

Main.Iot_Managed_Integrations.get_device_discovery — Function

get_device_discovery(identifier)
get_device_discovery(identifier, params::Dict{String,<:Any})

Get the current state of a device discovery.

Arguments

identifier: The id of the device discovery job request.

Main.Iot_Managed_Integrations.get_event_log_configuration — Function

get_event_log_configuration(id)
get_event_log_configuration(id, params::Dict{String,<:Any})

Get an event log configuration.

Arguments

id: The identifier of the event log configuration.

Main.Iot_Managed_Integrations.get_hub_configuration — Function

get_hub_configuration()
get_hub_configuration(params::Dict{String,<:Any})

Get a hub configuration.

Main.Iot_Managed_Integrations.get_managed_thing — Function

get_managed_thing(identifier)
get_managed_thing(identifier, params::Dict{String,<:Any})

Get details of a managed thing including its attributes and capabilities.

Arguments

identifier: The id of the managed thing.

Main.Iot_Managed_Integrations.get_managed_thing_capabilities — Function

get_managed_thing_capabilities(identifier)
get_managed_thing_capabilities(identifier, params::Dict{String,<:Any})

Get the capabilities for a managed thing using the device ID.

Arguments

identifier: The id of the device.

Main.Iot_Managed_Integrations.get_managed_thing_certificate — Function

get_managed_thing_certificate(identifier)
get_managed_thing_certificate(identifier, params::Dict{String,<:Any})

Retrieves the certificate PEM for a managed IoT thing.

Arguments

identifier: The identifier of the managed thing.

Main.Iot_Managed_Integrations.get_managed_thing_connectivity_data — Function

get_managed_thing_connectivity_data(identifier)
get_managed_thing_connectivity_data(identifier, params::Dict{String,<:Any})

Get the connectivity status of a managed thing.

Arguments

identifier: The identifier of a managed thing.

Main.Iot_Managed_Integrations.get_managed_thing_meta_data — Function

get_managed_thing_meta_data(identifier)
get_managed_thing_meta_data(identifier, params::Dict{String,<:Any})

Get the metadata information for a managed thing.

Note

The managedThing metadata parameter is used for associating attributes with a managedThing that can be used for grouping over-the-air (OTA) tasks. Name value pairs in metadata can be used in the OtaTargetQueryString parameter for the CreateOtaTask API operation.

Arguments

identifier: The managed thing id.

Main.Iot_Managed_Integrations.get_managed_thing_state — Function

get_managed_thing_state(managed_thing_id)
get_managed_thing_state(managed_thing_id, params::Dict{String,<:Any})

Returns the managed thing state for the given device Id.

Arguments

managed_thing_id: The id of the device.

Main.Iot_Managed_Integrations.get_notification_configuration — Function

get_notification_configuration(event_type)
get_notification_configuration(event_type, params::Dict{String,<:Any})

Get a notification configuration for a specified event type.

Arguments

event_type: The type of event triggering a device notification to the customer-managed destination.

Main.Iot_Managed_Integrations.get_ota_task — Function

get_ota_task(identifier)
get_ota_task(identifier, params::Dict{String,<:Any})

Get details of the over-the-air (OTA) task by its task id.

Arguments

identifier: The over-the-air (OTA) task id.

Main.Iot_Managed_Integrations.get_ota_task_configuration — Function

get_ota_task_configuration(identifier)
get_ota_task_configuration(identifier, params::Dict{String,<:Any})

Get a configuraiton for the over-the-air (OTA) task.

Arguments

identifier: The over-the-air (OTA) task configuration id.

Main.Iot_Managed_Integrations.get_provisioning_profile — Function

get_provisioning_profile(identifier)
get_provisioning_profile(identifier, params::Dict{String,<:Any})

Get a provisioning profile by template name.

Arguments

identifier: The provisioning template the device uses for the provisioning process.

Main.Iot_Managed_Integrations.get_runtime_log_configuration — Function

get_runtime_log_configuration(managed_thing_id)
get_runtime_log_configuration(managed_thing_id, params::Dict{String,<:Any})

Get the runtime log configuration for a specific managed thing.

Arguments

managed_thing_id: The id for a managed thing.

Main.Iot_Managed_Integrations.get_schema_version — Function

get_schema_version(schema_versioned_id, type)
get_schema_version(schema_versioned_id, type, params::Dict{String,<:Any})

Gets a schema version with the provided information.

Arguments

schema_versioned_id: Schema id with a version specified. If the version is missing, it defaults to latest version.

type: The type of schema version.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Format": The format of the schema version.

Main.Iot_Managed_Integrations.list_account_associations — Function

list_account_associations()
list_account_associations(params::Dict{String,<:Any})

Lists all account associations, with optional filtering by connector destination ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ConnectorDestinationId": The identifier of the connector destination to filter account associations by.

"MaxResults": The maximum number of account associations to return in a single response.

"NextToken": A token used for pagination of results.

Main.Iot_Managed_Integrations.list_cloud_connectors — Function

list_cloud_connectors()
list_cloud_connectors(params::Dict{String,<:Any})

Returns a list of connectors filtered by its Lambda Amazon Resource Name (ARN) and type.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"LambdaArn": The Amazon Resource Name (ARN) of the Lambda function to filter cloud connectors by.

"MaxResults": The maximum number of results to return at one time.

"NextToken": A token that can be used to retrieve the next set of results.

"Type": The type of cloud connectors to filter by when listing available connectors.

Main.Iot_Managed_Integrations.list_connector_destinations — Function

list_connector_destinations()
list_connector_destinations(params::Dict{String,<:Any})

Lists all connector destinations, with optional filtering by cloud connector ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"CloudConnectorId": The identifier of the cloud connector to filter connector destinations by.

"MaxResults": The maximum number of connector destinations to return in a single response.

"NextToken": A token used for pagination of results.

Main.Iot_Managed_Integrations.list_credential_lockers — Function

list_credential_lockers()
list_credential_lockers(params::Dict{String,<:Any})

List information on an existing credential locker.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return at one time.

"NextToken": A token that can be used to retrieve the next set of results.

Main.Iot_Managed_Integrations.list_destinations — Function

list_destinations()
list_destinations(params::Dict{String,<:Any})

List all notification destinations.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return at one time.

"NextToken": A token that can be used to retrieve the next set of results.

Main.Iot_Managed_Integrations.list_device_discoveries — Function

list_device_discoveries()
list_device_discoveries(params::Dict{String,<:Any})

Lists all device discovery tasks, with optional filtering by type and status.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of device discovery jobs to return in a single response.

"NextToken": A token used for pagination of results.

"StatusFilter": The status to filter device discovery jobs by.

"TypeFilter": The discovery type to filter device discovery jobs by.

Main.Iot_Managed_Integrations.list_discovered_devices — Function

list_discovered_devices(identifier)
list_discovered_devices(identifier, params::Dict{String,<:Any})

Lists all devices discovered during a specific device discovery task.

Arguments

identifier: The identifier of the device discovery job to list discovered devices for.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of discovered devices to return in a single response.

"NextToken": A token used for pagination of results.

Main.Iot_Managed_Integrations.list_event_log_configurations — Function

list_event_log_configurations()
list_event_log_configurations(params::Dict{String,<:Any})

List all event log configurations for an account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return at one time.

"NextToken": A token that can be used to retrieve the next set of results.

Main.Iot_Managed_Integrations.list_managed_thing_account_associations — Function

list_managed_thing_account_associations()
list_managed_thing_account_associations(params::Dict{String,<:Any})

Lists all account associations for a specific managed thing.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AccountAssociationId": The identifier of the account association to filter results by. When specified, only associations with this account association ID will be returned.

"ManagedThingId": The identifier of the managed thing to list account associations for.

"MaxResults": The maximum number of account associations to return in a single response.

"NextToken": A token used for pagination of results.

Main.Iot_Managed_Integrations.list_managed_thing_schemas — Function

list_managed_thing_schemas(identifier)
list_managed_thing_schemas(identifier, params::Dict{String,<:Any})

List schemas associated with a managed thing.

Arguments

identifier: The managed thing id.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"CapabilityIdFilter": Filter on a capability id.

"EndpointIdFilter": Filter on an endpoint id.

"MaxResults": The maximum number of results to return at one time.

"NextToken": A token that can be used to retrieve the next set of results.

Main.Iot_Managed_Integrations.list_managed_things — Function

list_managed_things()
list_managed_things(params::Dict{String,<:Any})

Listing all managed things with provision for filters.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ConnectorDestinationIdFilter": Filter managed things by the connector destination ID they are associated with.

"ConnectorDeviceIdFilter": Filter managed things by the connector device ID they are associated with. When specified, only managed things with this connector device ID will be returned.

"ConnectorPolicyIdFilter": Filter on a connector policy id for a managed thing.

"CredentialLockerFilter": Filter on a credential locker for a managed thing.

"MaxResults": The maximum number of results to return at one time.

"NextToken": A token that can be used to retrieve the next set of results.

"OwnerFilter": Filter on device owners when listing managed things.

"ParentControllerIdentifierFilter": Filter on a parent controller id for a managed thing.

"ProvisioningStatusFilter": Filter on the status of the device. For more information, see Device Provisioning.

"RoleFilter": Filter on the type of device used. This will be the Amazon Web Services hub controller, cloud device, or IoT device.

"SerialNumberFilter": Filter on the serial number of the device.

Main.Iot_Managed_Integrations.list_notification_configurations — Function

list_notification_configurations()
list_notification_configurations(params::Dict{String,<:Any})

List all notification configurations.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return at one time.

"NextToken": A token that can be used to retrieve the next set of results.

Main.Iot_Managed_Integrations.list_ota_task_configurations — Function

list_ota_task_configurations()
list_ota_task_configurations(params::Dict{String,<:Any})

List all of the over-the-air (OTA) task configurations.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return at one time.

"NextToken": A token that can be used to retrieve the next set of results.

Main.Iot_Managed_Integrations.list_ota_task_executions — Function

list_ota_task_executions(identifier)
list_ota_task_executions(identifier, params::Dict{String,<:Any})

List all of the over-the-air (OTA) task executions.

Arguments

identifier: The over-the-air (OTA) task id.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return at one time.

"NextToken": A token that can be used to retrieve the next set of results.

Main.Iot_Managed_Integrations.list_ota_tasks — Function

list_ota_tasks()
list_ota_tasks(params::Dict{String,<:Any})

List all of the over-the-air (OTA) tasks.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return at one time.

"NextToken": A token that can be used to retrieve the next set of results.

Main.Iot_Managed_Integrations.list_provisioning_profiles — Function

list_provisioning_profiles()
list_provisioning_profiles(params::Dict{String,<:Any})

List the provisioning profiles within the Amazon Web Services account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return at one time.

"NextToken": A token that can be used to retrieve the next set of results.

Main.Iot_Managed_Integrations.list_schema_versions — Function

list_schema_versions(type)
list_schema_versions(type, params::Dict{String,<:Any})

Lists schema versions with the provided information.

Arguments

type: Filter on the type of schema version.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return at one time.

"NamespaceFilter": Filter on the name of the schema version.

"NextToken": A token that can be used to retrieve the next set of results.

"SchemaIdFilter": Filter on the id of the schema version.

"SemanticVersionFilter": The schema version. If this is left blank, it defaults to the latest version.

"VisibilityFilter": The visibility of the schema version.

Main.Iot_Managed_Integrations.list_tags_for_resource — Function

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists the tags for a specified resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource for which to list tags.

Main.Iot_Managed_Integrations.put_default_encryption_configuration — Function

put_default_encryption_configuration(encryption_type)
put_default_encryption_configuration(encryption_type, params::Dict{String,<:Any})

Sets the default encryption configuration for the Amazon Web Services account. For more information, see Key management in the AWS IoT SiteWise User Guide.

Arguments

encryption_type: The type of encryption used for the encryption configuration.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"kmsKeyArn": The Key Amazon Resource Name (ARN) of the AWS KMS key used for KMS encryption if you use KMS_BASED_ENCRYPTION.

Main.Iot_Managed_Integrations.put_hub_configuration — Function

put_hub_configuration(hub_token_timer_expiry_setting_in_seconds)
put_hub_configuration(hub_token_timer_expiry_setting_in_seconds, params::Dict{String,<:Any})

Update a hub configuration.

Arguments

hub_token_timer_expiry_setting_in_seconds: A user-defined integer value that represents the hub token timer expiry setting in seconds.

Main.Iot_Managed_Integrations.put_runtime_log_configuration — Function

put_runtime_log_configuration(managed_thing_id, runtime_log_configurations)
put_runtime_log_configuration(managed_thing_id, runtime_log_configurations, params::Dict{String,<:Any})

Set the runtime log configuration for a specific managed thing.

Arguments

managed_thing_id: The id for a managed thing.

runtime_log_configurations: The runtime log configuration for a managed thing.

Main.Iot_Managed_Integrations.register_account_association — Function

register_account_association(account_association_id, device_discovery_id, managed_thing_id)
register_account_association(account_association_id, device_discovery_id, managed_thing_id, params::Dict{String,<:Any})

Registers an account association with a managed thing, establishing a connection between a device and a third-party account.

Arguments

account_association_id: The identifier of the account association to register with the managed thing.

device_discovery_id: The identifier of the device discovery job associated with this registration.

managed_thing_id: The identifier of the managed thing to register with the account association.

Main.Iot_Managed_Integrations.register_custom_endpoint — Function

register_custom_endpoint()
register_custom_endpoint(params::Dict{String,<:Any})

Customers can request IoT managed integrations to manage the server trust for them or bring their own external server trusts for the custom domain. Returns an IoT managed integrations endpoint.

Main.Iot_Managed_Integrations.reset_runtime_log_configuration — Function

reset_runtime_log_configuration(managed_thing_id)
reset_runtime_log_configuration(managed_thing_id, params::Dict{String,<:Any})

Reset a runtime log configuration for a specific managed thing.

Arguments

managed_thing_id: The id of a managed thing.

Main.Iot_Managed_Integrations.send_connector_event — Function

send_connector_event(connector_id, operation)
send_connector_event(connector_id, operation, params::Dict{String,<:Any})

Relays third-party device events for a connector such as a new device or a device state change event.

Arguments

connector_id: The id of the connector between the third-party cloud provider and IoT managed integrations.

operation: The Open Connectivity Foundation (OCF) operation requested to be performed on the managed thing.

Note

The field op can have a value of "I" or "U". The field "cn" will contain the capability types.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ConnectorDeviceId": The third-party device id as defined by the connector. This device id must not contain personal identifiable information (PII).

Note

This parameter is used for cloud-to-cloud devices only.

"DeviceDiscoveryId": The id for the device discovery job.

"Devices": The list of devices.

"MatterEndpoint": The device endpoint.

"Message": The device state change event payload.

This parameter will include the following three fields:

uri: schema auc://&lt;PARTNER-DEVICE-ID&gt;/ResourcePath (The Resourcepath corresponds to an OCF resource.)

op: For device state changes, this field must populate as n+d.

cn: The content depends on the OCF resource referenced in ResourcePath.

"OperationVersion": The Open Connectivity Foundation (OCF) security specification version for the operation being requested on the managed thing. For more information, see OCF Security Specification.

"StatusCode": The status code of the Open Connectivity Foundation (OCF) operation being performed on the managed thing.

"TraceId": The trace request identifier. This is generated by IoT managed integrations and can be used to trace this command and its related operations in CloudWatch.

"UserId": The id of the third-party cloud provider.

Main.Iot_Managed_Integrations.send_managed_thing_command — Function

send_managed_thing_command(endpoints, managed_thing_id)
send_managed_thing_command(endpoints, managed_thing_id, params::Dict{String,<:Any})

Send the command to the device represented by the managed thing.

Arguments

endpoints: The device endpoint.

managed_thing_id: The id of the device.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AccountAssociationId": The identifier of the account association to use when sending a command to a managed thing.

"ConnectorAssociationId": The ID tracking the current discovery process for one connector association.

Main.Iot_Managed_Integrations.start_account_association_refresh — Function

start_account_association_refresh(account_association_id)
start_account_association_refresh(account_association_id, params::Dict{String,<:Any})

Initiates a refresh of an existing account association to update its authorization and connection status.

Arguments

account_association_id: The unique identifier of the account association to refresh.

Main.Iot_Managed_Integrations.start_device_discovery — Function

start_device_discovery(discovery_type)
start_device_discovery(discovery_type, params::Dict{String,<:Any})

This API is used to start device discovery for hub-connected and third-party-connected devices. The authentication material (install code) is delivered as a message to the controller instructing it to start the discovery.

Arguments

discovery_type: The discovery type supporting the type of device to be discovered in the device discovery task request.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AccountAssociationId": The identifier of the cloud-to-cloud account association to use for discovery of third-party devices.

"AuthenticationMaterial": The authentication material required to start the local device discovery job request.

"AuthenticationMaterialType": The type of authentication material used for device discovery jobs.

"ClientToken": An idempotency token. If you retry a request that completed successfully initially using the same client token and parameters, then the retry attempt will succeed without performing any further actions.

"ConnectorAssociationIdentifier": The id of the connector association.

"ConnectorDeviceIdList": Used as a filter for PLA discoveries.

"ControllerIdentifier": The id of the end-user's IoT hub.

"CustomProtocolDetail": Additional protocol-specific details required for device discovery, which vary based on the discovery type.

Note

For a DiscoveryType of CUSTOM, the string-to-string map must have a key value of Name set to a non-empty-string.

"EndDeviceIdentifier": The unique id of the end device for capability rediscovery.

Note

This parameter is only available when the discovery type is CONTROLLERCAPABILITYREDISCOVERY.

"Protocol": The protocol type for capability rediscovery (ZWAVE, ZIGBEE, or CUSTOM).

Note

This parameter is only available when the discovery type is CONTROLLERCAPABILITYREDISCOVERY.

"Tags": A set of key/value pairs that are used to manage the device discovery request.

Main.Iot_Managed_Integrations.tag_resource — Function

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Adds tags to a specified resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource to which to add tags.

tags: A set of key/value pairs that are used to manage the resource.

Main.Iot_Managed_Integrations.untag_resource — Function

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes tags from a specified resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource from which to remove tags.

tag_keys: A list of tag keys to remove from the resource.

Main.Iot_Managed_Integrations.update_account_association — Function

update_account_association(account_association_id)
update_account_association(account_association_id, params::Dict{String,<:Any})

Updates the properties of an existing account association.

Arguments

account_association_id: The unique identifier of the account association to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": The new description to assign to the account association.

"Name": The new name to assign to the account association.

Main.Iot_Managed_Integrations.update_cloud_connector — Function

update_cloud_connector(identifier)
update_cloud_connector(identifier, params::Dict{String,<:Any})

Update an existing cloud connector.

Arguments

identifier: The unique identifier of the cloud connector to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": The new description to assign to the cloud connector.

"Name": The new display name to assign to the cloud connector.

Main.Iot_Managed_Integrations.update_connector_destination — Function

update_connector_destination(identifier)
update_connector_destination(identifier, params::Dict{String,<:Any})

Updates the properties of an existing connector destination.

Arguments

identifier: The unique identifier of the connector destination to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AuthConfig": The updated authentication configuration details for the connector destination.

"AuthType": The new authentication type to use for the connector destination.

"Description": The new description to assign to the connector destination.

"Name": The new display name to assign to the connector destination.

"SecretsManager": The updated AWS Secrets Manager configuration for the connector destination.

Main.Iot_Managed_Integrations.update_destination — Function

update_destination(name)
update_destination(name, params::Dict{String,<:Any})

Update a destination specified by name.

Arguments

name: The name of the customer-managed destination.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DeliveryDestinationArn": The Amazon Resource Name (ARN) of the customer-managed destination.

"DeliveryDestinationType": The destination type for the customer-managed destination.

"Description": The description of the customer-managed destination.

"RoleArn": The Amazon Resource Name (ARN) of the delivery destination role.

Main.Iot_Managed_Integrations.update_event_log_configuration — Function

update_event_log_configuration(event_log_level, id)
update_event_log_configuration(event_log_level, id, params::Dict{String,<:Any})

Update an event log configuration by log configuration ID.

Arguments

event_log_level: The log level for the event in terms of severity.

id: The log configuration id.

Main.Iot_Managed_Integrations.update_managed_thing — Function

update_managed_thing(identifier)
update_managed_thing(identifier, params::Dict{String,<:Any})

Update the attributes and capabilities associated with a managed thing.

Arguments

identifier: The id of the managed thing.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Brand": The brand of the device.

"Capabilities": The capabilities of the device such as light bulb.

"CapabilityReport": A report of the capabilities for the managed thing.

"CapabilitySchemas": The updated capability schemas that define the functionality and features supported by the managed thing.

"Classification": The classification of the managed thing such as light bulb or thermostat.

"CredentialLockerId": The identifier of the credential for the managed thing.

"HubNetworkMode": The network mode for the hub-connected device.

"MetaData": The metadata for the managed thing.

"Model": The model of the device.

"Name": The name of the managed thing representing the physical device.

"Owner": Owner of the device, usually an indication of whom the device belongs to. This value should not contain personal identifiable information.

"SerialNumber": The serial number of the device.

"WiFiSimpleSetupConfiguration": The Wi-Fi Simple Setup configuration for the managed thing, which defines provisioning capabilities and timeout settings.

Main.Iot_Managed_Integrations.update_notification_configuration — Function

update_notification_configuration(destination_name, event_type)
update_notification_configuration(destination_name, event_type, params::Dict{String,<:Any})

Update a notification configuration.

Arguments

destination_name: The name of the destination for the notification configuration.

event_type: The type of event triggering a device notification to the customer-managed destination.

Main.Iot_Managed_Integrations.update_ota_task — Function

update_ota_task(identifier)
update_ota_task(identifier, params::Dict{String,<:Any})

Update an over-the-air (OTA) task.

Arguments

identifier: The over-the-air (OTA) task id.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": The description of the over-the-air (OTA) task.

"TaskConfigurationId": The identifier for the over-the-air (OTA) task configuration.