Iot Wireless

This page documents function available when using the Iot_Wireless module, created with @service Iot_Wireless.

Index

Main.Iot_Wireless.associate_aws_account_with_partner_account

Main.Iot_Wireless.associate_wireless_device_with_thing

Main.Iot_Wireless.associate_wireless_gateway_with_certificate

Main.Iot_Wireless.associate_wireless_gateway_with_thing

Main.Iot_Wireless.create_destination

Main.Iot_Wireless.create_device_profile

Main.Iot_Wireless.create_service_profile

Main.Iot_Wireless.create_wireless_device

Main.Iot_Wireless.create_wireless_gateway

Main.Iot_Wireless.create_wireless_gateway_task

Main.Iot_Wireless.create_wireless_gateway_task_definition

Main.Iot_Wireless.delete_destination

Main.Iot_Wireless.delete_device_profile

Main.Iot_Wireless.delete_service_profile

Main.Iot_Wireless.delete_wireless_device

Main.Iot_Wireless.delete_wireless_gateway

Main.Iot_Wireless.delete_wireless_gateway_task

Main.Iot_Wireless.delete_wireless_gateway_task_definition

Main.Iot_Wireless.disassociate_aws_account_from_partner_account

Main.Iot_Wireless.disassociate_wireless_device_from_thing

Main.Iot_Wireless.disassociate_wireless_gateway_from_certificate

Main.Iot_Wireless.disassociate_wireless_gateway_from_thing

Main.Iot_Wireless.get_destination

Main.Iot_Wireless.get_device_profile

Main.Iot_Wireless.get_partner_account

Main.Iot_Wireless.get_service_endpoint

Main.Iot_Wireless.get_service_profile

Main.Iot_Wireless.get_wireless_device

Main.Iot_Wireless.get_wireless_device_statistics

Main.Iot_Wireless.get_wireless_gateway

Main.Iot_Wireless.get_wireless_gateway_certificate

Main.Iot_Wireless.get_wireless_gateway_firmware_information

Main.Iot_Wireless.get_wireless_gateway_statistics

Main.Iot_Wireless.get_wireless_gateway_task

Main.Iot_Wireless.get_wireless_gateway_task_definition

Main.Iot_Wireless.list_destinations

Main.Iot_Wireless.list_device_profiles

Main.Iot_Wireless.list_partner_accounts

Main.Iot_Wireless.list_service_profiles

Main.Iot_Wireless.list_tags_for_resource

Main.Iot_Wireless.list_wireless_devices

Main.Iot_Wireless.list_wireless_gateway_task_definitions

Main.Iot_Wireless.list_wireless_gateways

Main.Iot_Wireless.send_data_to_wireless_device

Main.Iot_Wireless.tag_resource

Main.Iot_Wireless.test_wireless_device

Main.Iot_Wireless.untag_resource

Main.Iot_Wireless.update_destination

Main.Iot_Wireless.update_partner_account

Main.Iot_Wireless.update_wireless_device

Main.Iot_Wireless.update_wireless_gateway

Documentation

Main.Iot_Wireless.associate_aws_account_with_partner_account — Method

associate_aws_account_with_partner_account(sidewalk)
associate_aws_account_with_partner_account(sidewalk, params::Dict{String,<:Any})

Associates a partner account with your AWS account.

Arguments

sidewalk: The Sidewalk account credentials.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientRequestToken": Each resource must have a unique client request token. If you try to create a new resource with the same token as a resource that already exists, an exception occurs. If you omit this value, AWS SDKs will automatically generate a unique client request.

"Tags": The tags to attach to the specified resource. Tags are metadata that you can use to manage a resource.

Main.Iot_Wireless.associate_wireless_device_with_thing — Method

associate_wireless_device_with_thing(id, thing_arn)
associate_wireless_device_with_thing(id, thing_arn, params::Dict{String,<:Any})

Associates a wireless device with a thing.

Arguments

id: The ID of the resource to update.

thing_arn: The ARN of the thing to associate with the wireless device.

Main.Iot_Wireless.associate_wireless_gateway_with_certificate — Method

associate_wireless_gateway_with_certificate(id, iot_certificate_id)
associate_wireless_gateway_with_certificate(id, iot_certificate_id, params::Dict{String,<:Any})

Associates a wireless gateway with a certificate.

Arguments

id: The ID of the resource to update.

iot_certificate_id: The ID of the certificate to associate with the wireless gateway.

Main.Iot_Wireless.associate_wireless_gateway_with_thing — Method

associate_wireless_gateway_with_thing(id, thing_arn)
associate_wireless_gateway_with_thing(id, thing_arn, params::Dict{String,<:Any})

Associates a wireless gateway with a thing.

Arguments

id: The ID of the resource to update.

thing_arn: The ARN of the thing to associate with the wireless gateway.

Main.Iot_Wireless.create_destination — Method

create_destination(expression, expression_type, name, role_arn)
create_destination(expression, expression_type, name, role_arn, params::Dict{String,<:Any})

Creates a new destination that maps a device message to an AWS IoT rule.

Arguments

expression: The rule name or topic rule to send messages to.

expression_type: The type of value in Expression.

name: The name of the new resource.

role_arn: The ARN of the IAM Role that authorizes the destination.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientRequestToken": Each resource must have a unique client request token. If you try to create a new resource with the same token as a resource that already exists, an exception occurs. If you omit this value, AWS SDKs will automatically generate a unique client request.

"Description": The description of the new resource.

"Tags": The tags to attach to the new destination. Tags are metadata that you can use to manage a resource.

Main.Iot_Wireless.create_device_profile — Method

create_device_profile()
create_device_profile(params::Dict{String,<:Any})

Creates a new device profile.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientRequestToken": Each resource must have a unique client request token. If you try to create a new resource with the same token as a resource that already exists, an exception occurs. If you omit this value, AWS SDKs will automatically generate a unique client request.

"LoRaWAN": The device profile information to use to create the device profile.

"Name": The name of the new resource.

"Tags": The tags to attach to the new device profile. Tags are metadata that you can use to manage a resource.

Main.Iot_Wireless.create_service_profile — Method

create_service_profile()
create_service_profile(params::Dict{String,<:Any})

Creates a new service profile.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientRequestToken": Each resource must have a unique client request token. If you try to create a new resource with the same token as a resource that already exists, an exception occurs. If you omit this value, AWS SDKs will automatically generate a unique client request.

"LoRaWAN": The service profile information to use to create the service profile.

"Name": The name of the new resource.

"Tags": The tags to attach to the new service profile. Tags are metadata that you can use to manage a resource.

Main.Iot_Wireless.create_wireless_device — Method

create_wireless_device(destination_name, type)
create_wireless_device(destination_name, type, params::Dict{String,<:Any})

Provisions a wireless device.

Arguments

destination_name: The name of the destination to assign to the new wireless device.

type: The wireless device type.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientRequestToken": Each resource must have a unique client request token. If you try to create a new resource with the same token as a resource that already exists, an exception occurs. If you omit this value, AWS SDKs will automatically generate a unique client request.

"Description": The description of the new resource.

"LoRaWAN": The device configuration information to use to create the wireless device.

"Name": The name of the new resource.

"Tags": The tags to attach to the new wireless device. Tags are metadata that you can use to manage a resource.

Main.Iot_Wireless.create_wireless_gateway — Method

create_wireless_gateway(lo_ra_wan)
create_wireless_gateway(lo_ra_wan, params::Dict{String,<:Any})

Provisions a wireless gateway.

Arguments

lo_ra_wan: The gateway configuration information to use to create the wireless gateway.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientRequestToken": Each resource must have a unique client request token. If you try to create a new resource with the same token as a resource that already exists, an exception occurs. If you omit this value, AWS SDKs will automatically generate a unique client request.

"Description": The description of the new resource.

"Name": The name of the new resource.

"Tags": The tags to attach to the new wireless gateway. Tags are metadata that you can use to manage a resource.

Main.Iot_Wireless.create_wireless_gateway_task — Method

create_wireless_gateway_task(id, wireless_gateway_task_definition_id)
create_wireless_gateway_task(id, wireless_gateway_task_definition_id, params::Dict{String,<:Any})

Creates a task for a wireless gateway.

Arguments

id: The ID of the resource to update.

wireless_gateway_task_definition_id: The ID of the WirelessGatewayTaskDefinition.

Main.Iot_Wireless.create_wireless_gateway_task_definition — Method

create_wireless_gateway_task_definition(auto_create_tasks)
create_wireless_gateway_task_definition(auto_create_tasks, params::Dict{String,<:Any})

Creates a gateway task definition.

Arguments

auto_create_tasks: Whether to automatically create tasks using this task definition for all gateways with the specified current version. If false, the task must me created by calling CreateWirelessGatewayTask.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientRequestToken": Each resource must have a unique client request token. If you try to create a new resource with the same token as a resource that already exists, an exception occurs. If you omit this value, AWS SDKs will automatically generate a unique client request.

"Name": The name of the new resource.

"Tags": The tags to attach to the specified resource. Tags are metadata that you can use to manage a resource.

"Update": Information about the gateways to update.

Main.Iot_Wireless.delete_destination — Method

delete_destination(name)
delete_destination(name, params::Dict{String,<:Any})

Deletes a destination.

Arguments

name: The name of the resource to delete.

Main.Iot_Wireless.delete_device_profile — Method

delete_device_profile(id)
delete_device_profile(id, params::Dict{String,<:Any})

Deletes a device profile.

Arguments

id: The ID of the resource to delete.

Main.Iot_Wireless.delete_service_profile — Method

delete_service_profile(id)
delete_service_profile(id, params::Dict{String,<:Any})

Deletes a service profile.

Arguments

id: The ID of the resource to delete.

Main.Iot_Wireless.delete_wireless_device — Method

delete_wireless_device(id)
delete_wireless_device(id, params::Dict{String,<:Any})

Deletes a wireless device.

Arguments

id: The ID of the resource to delete.

Main.Iot_Wireless.delete_wireless_gateway — Method

delete_wireless_gateway(id)
delete_wireless_gateway(id, params::Dict{String,<:Any})

Deletes a wireless gateway.

Arguments

id: The ID of the resource to delete.

Main.Iot_Wireless.delete_wireless_gateway_task — Method

delete_wireless_gateway_task(id)
delete_wireless_gateway_task(id, params::Dict{String,<:Any})

Deletes a wireless gateway task.

Arguments

id: The ID of the resource to delete.

Main.Iot_Wireless.delete_wireless_gateway_task_definition — Method

delete_wireless_gateway_task_definition(id)
delete_wireless_gateway_task_definition(id, params::Dict{String,<:Any})

Deletes a wireless gateway task definition. Deleting this task definition does not affect tasks that are currently in progress.

Arguments

id: The ID of the resource to delete.

Main.Iot_Wireless.disassociate_aws_account_from_partner_account — Method

disassociate_aws_account_from_partner_account(partner_account_id, partner_type)
disassociate_aws_account_from_partner_account(partner_account_id, partner_type, params::Dict{String,<:Any})

Disassociates your AWS account from a partner account. If PartnerAccountId and PartnerType are null, disassociates your AWS account from all partner accounts.

Arguments

partner_account_id: The partner account ID to disassociate from the AWS account.

partner_type: The partner type.

Main.Iot_Wireless.disassociate_wireless_device_from_thing — Method

disassociate_wireless_device_from_thing(id)
disassociate_wireless_device_from_thing(id, params::Dict{String,<:Any})

Disassociates a wireless device from its currently associated thing.

Arguments

id: The ID of the resource to update.

Main.Iot_Wireless.disassociate_wireless_gateway_from_certificate — Method

disassociate_wireless_gateway_from_certificate(id)
disassociate_wireless_gateway_from_certificate(id, params::Dict{String,<:Any})

Disassociates a wireless gateway from its currently associated certificate.

Arguments

id: The ID of the resource to update.

Main.Iot_Wireless.disassociate_wireless_gateway_from_thing — Method

disassociate_wireless_gateway_from_thing(id)
disassociate_wireless_gateway_from_thing(id, params::Dict{String,<:Any})

Disassociates a wireless gateway from its currently associated thing.

Arguments

id: The ID of the resource to update.

Main.Iot_Wireless.get_destination — Method

get_destination(name)
get_destination(name, params::Dict{String,<:Any})

Gets information about a destination.

Arguments

name: The name of the resource to get.

Main.Iot_Wireless.get_device_profile — Method

get_device_profile(id)
get_device_profile(id, params::Dict{String,<:Any})

Gets information about a device profile.

Arguments

id: The ID of the resource to get.

Main.Iot_Wireless.get_partner_account — Method

get_partner_account(partner_account_id, partner_type)
get_partner_account(partner_account_id, partner_type, params::Dict{String,<:Any})

Gets information about a partner account. If PartnerAccountId and PartnerType are null, returns all partner accounts.

Arguments

partner_account_id: The partner account ID to disassociate from the AWS account.

partner_type: The partner type.

Main.Iot_Wireless.get_service_endpoint — Method

get_service_endpoint()
get_service_endpoint(params::Dict{String,<:Any})

Gets the account-specific endpoint for Configuration and Update Server (CUPS) protocol or LoRaWAN Network Server (LNS) connections.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"serviceType": The service type for which to get endpoint information about. Can be CUPS for the Configuration and Update Server endpoint, or LNS for the LoRaWAN Network Server endpoint.

Main.Iot_Wireless.get_service_profile — Method

get_service_profile(id)
get_service_profile(id, params::Dict{String,<:Any})

Gets information about a service profile.

Arguments

id: The ID of the resource to get.

Main.Iot_Wireless.get_wireless_device — Method

get_wireless_device(identifier, identifier_type)
get_wireless_device(identifier, identifier_type, params::Dict{String,<:Any})

Gets information about a wireless device.

Arguments

identifier: The identifier of the wireless device to get.

identifier_type: The type of identifier used in identifier.

Main.Iot_Wireless.get_wireless_device_statistics — Method

get_wireless_device_statistics(id)
get_wireless_device_statistics(id, params::Dict{String,<:Any})

Gets operating information about a wireless device.

Arguments

id: The ID of the wireless device for which to get the data.

Main.Iot_Wireless.get_wireless_gateway — Method

get_wireless_gateway(identifier, identifier_type)
get_wireless_gateway(identifier, identifier_type, params::Dict{String,<:Any})

Gets information about a wireless gateway.

Arguments

identifier: The identifier of the wireless gateway to get.

identifier_type: The type of identifier used in identifier.

Main.Iot_Wireless.get_wireless_gateway_certificate — Method

get_wireless_gateway_certificate(id)
get_wireless_gateway_certificate(id, params::Dict{String,<:Any})

Gets the ID of the certificate that is currently associated with a wireless gateway.

Arguments

id: The ID of the resource to get.

Main.Iot_Wireless.get_wireless_gateway_firmware_information — Method

get_wireless_gateway_firmware_information(id)
get_wireless_gateway_firmware_information(id, params::Dict{String,<:Any})

Gets the firmware version and other information about a wireless gateway.

Arguments

id: The ID of the resource to get.

Main.Iot_Wireless.get_wireless_gateway_statistics — Method

get_wireless_gateway_statistics(id)
get_wireless_gateway_statistics(id, params::Dict{String,<:Any})

Gets operating information about a wireless gateway.

Arguments

id: The ID of the wireless gateway for which to get the data.

Main.Iot_Wireless.get_wireless_gateway_task — Method

get_wireless_gateway_task(id)
get_wireless_gateway_task(id, params::Dict{String,<:Any})

Gets information about a wireless gateway task.

Arguments

id: The ID of the resource to get.

Main.Iot_Wireless.get_wireless_gateway_task_definition — Method

get_wireless_gateway_task_definition(id)
get_wireless_gateway_task_definition(id, params::Dict{String,<:Any})

Gets information about a wireless gateway task definition.

Arguments

id: The ID of the resource to get.

Main.Iot_Wireless.list_destinations — Method

list_destinations()
list_destinations(params::Dict{String,<:Any})

Lists the destinations registered to your AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return in this operation.

"nextToken": To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

Main.Iot_Wireless.list_device_profiles — Method

list_device_profiles()
list_device_profiles(params::Dict{String,<:Any})

Lists the device profiles registered to your AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return in this operation.

"nextToken": To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

Main.Iot_Wireless.list_partner_accounts — Method

list_partner_accounts()
list_partner_accounts(params::Dict{String,<:Any})

Lists the partner accounts associated with your AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return in this operation.

"nextToken": To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

Main.Iot_Wireless.list_service_profiles — Method

list_service_profiles()
list_service_profiles(params::Dict{String,<:Any})

Lists the service profiles registered to your AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return in this operation.

"nextToken": To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

Main.Iot_Wireless.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists the tags (metadata) you have assigned to the resource.

Arguments

resource_arn: The ARN of the resource for which you want to list tags.

Main.Iot_Wireless.list_wireless_devices — Method

list_wireless_devices()
list_wireless_devices(params::Dict{String,<:Any})

Lists the wireless devices registered to your AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"destinationName": A filter to list only the wireless devices that use this destination.

"deviceProfileId": A filter to list only the wireless devices that use this device profile.

"maxResults": The maximum number of results to return in this operation.

"nextToken": To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

"serviceProfileId": A filter to list only the wireless devices that use this service profile.

"wirelessDeviceType": A filter to list only the wireless devices that use this wireless device type.

Main.Iot_Wireless.list_wireless_gateway_task_definitions — Method

list_wireless_gateway_task_definitions()
list_wireless_gateway_task_definitions(params::Dict{String,<:Any})

List the wireless gateway tasks definitions registered to your AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return in this operation.

"nextToken": To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

"taskDefinitionType": A filter to list only the wireless gateway task definitions that use this task definition type.

Main.Iot_Wireless.list_wireless_gateways — Method

list_wireless_gateways()
list_wireless_gateways(params::Dict{String,<:Any})

Lists the wireless gateways registered to your AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return in this operation.

"nextToken": To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

Main.Iot_Wireless.send_data_to_wireless_device — Method

send_data_to_wireless_device(id, payload_data, transmit_mode)
send_data_to_wireless_device(id, payload_data, transmit_mode, params::Dict{String,<:Any})

Sends a decrypted application data frame to a device.

Arguments

id: The ID of the wireless device to receive the data.

payload_data: The message payload to send.

transmit_mode: The transmit mode to use to send data to the wireless device. Can be: 0 for UM (unacknowledge mode) or 1 for AM (acknowledge mode).

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"WirelessMetadata": Metadata about the message request.

Main.Iot_Wireless.tag_resource — Method

tag_resource(tags, resource_arn)
tag_resource(tags, resource_arn, params::Dict{String,<:Any})

Adds a tag to a resource.

Arguments

tags: Adds to or modifies the tags of the given resource. Tags are metadata that you can use to manage a resource.

resource_arn: The ARN of the resource to add tags to.

Main.Iot_Wireless.test_wireless_device — Method

test_wireless_device(id)
test_wireless_device(id, params::Dict{String,<:Any})

Simulates a provisioned device by sending an uplink data payload of Hello.

Arguments

id: The ID of the wireless device to test.

Main.Iot_Wireless.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes one or more tags from a resource.

Arguments

resource_arn: The ARN of the resource to remove tags from.

tag_keys: A list of the keys of the tags to remove from the resource.

Main.Iot_Wireless.update_destination — Method

update_destination(name)
update_destination(name, params::Dict{String,<:Any})

Updates properties of a destination.

Arguments

name: The new name of the resource.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A new description of the resource.

"Expression": The new rule name or topic rule to send messages to.

"ExpressionType": The type of value in Expression.

"RoleArn": The ARN of the IAM Role that authorizes the destination.

Main.Iot_Wireless.update_partner_account — Method

update_partner_account(partner_account_id, sidewalk, partner_type)
update_partner_account(partner_account_id, sidewalk, partner_type, params::Dict{String,<:Any})

Updates properties of a partner account.

Arguments

partner_account_id: The ID of the partner account to update.

sidewalk: The Sidewalk account credentials.

partner_type: The partner type.

Main.Iot_Wireless.update_wireless_device — Method

update_wireless_device(id)
update_wireless_device(id, params::Dict{String,<:Any})

Updates properties of a wireless device.

Arguments

id: The ID of the resource to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A new description of the resource.

"DestinationName": The name of the new destination for the device.

"LoRaWAN": The updated wireless device's configuration.

"Name": The new name of the resource.

Main.Iot_Wireless.update_wireless_gateway — Method

update_wireless_gateway(id)
update_wireless_gateway(id, params::Dict{String,<:Any})

Updates properties of a wireless gateway.

Arguments

id: The ID of the resource to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A new description of the resource.

"Name": The new name of the resource.