Kafkaconnect

This page documents function available when using the Kafkaconnect module, created with @service Kafkaconnect.

Index

Main.Kafkaconnect.create_connector

Main.Kafkaconnect.create_custom_plugin

Main.Kafkaconnect.create_worker_configuration

Main.Kafkaconnect.delete_connector

Main.Kafkaconnect.describe_connector

Main.Kafkaconnect.describe_custom_plugin

Main.Kafkaconnect.describe_worker_configuration

Main.Kafkaconnect.list_connectors

Main.Kafkaconnect.list_custom_plugins

Main.Kafkaconnect.list_worker_configurations

Main.Kafkaconnect.update_connector

Documentation

Main.Kafkaconnect.create_connector — Method

create_connector(capacity, connector_configuration, connector_name, kafka_cluster, kafka_cluster_client_authentication, kafka_cluster_encryption_in_transit, kafka_connect_version, plugins, service_execution_role_arn)
create_connector(capacity, connector_configuration, connector_name, kafka_cluster, kafka_cluster_client_authentication, kafka_cluster_encryption_in_transit, kafka_connect_version, plugins, service_execution_role_arn, params::Dict{String,<:Any})

Creates a connector using the specified properties.

Arguments

capacity: Information about the capacity allocated to the connector. Exactly one of the two properties must be specified.

connector_configuration: A map of keys to values that represent the configuration for the connector.

connector_name: The name of the connector.

kafka_cluster: Specifies which Apache Kafka cluster to connect to.

kafka_cluster_client_authentication: Details of the client authentication used by the Apache Kafka cluster.

kafka_cluster_encryption_in_transit: Details of encryption in transit to the Apache Kafka cluster.

kafka_connect_version: The version of Kafka Connect. It has to be compatible with both the Apache Kafka cluster's version and the plugins.

plugins: Specifies which plugins to use for the connector.

service_execution_role_arn: The Amazon Resource Name (ARN) of the IAM role used by the connector to access the Amazon Web Services resources that it needs. The types of resources depends on the logic of the connector. For example, a connector that has Amazon S3 as a destination must have permissions that allow it to write to the S3 destination bucket.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"connectorDescription": A summary description of the connector.

"logDelivery": Details about log delivery.

"workerConfiguration": Specifies which worker configuration to use with the connector.

Main.Kafkaconnect.create_custom_plugin — Method

create_custom_plugin(content_type, location, name)
create_custom_plugin(content_type, location, name, params::Dict{String,<:Any})

Creates a custom plugin using the specified properties.

Arguments

content_type: The type of the plugin file.

location: Information about the location of a custom plugin.

name: The name of the custom plugin.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A summary description of the custom plugin.

Main.Kafkaconnect.create_worker_configuration — Method

create_worker_configuration(name, properties_file_content)
create_worker_configuration(name, properties_file_content, params::Dict{String,<:Any})

Creates a worker configuration using the specified properties.

Arguments

name: The name of the worker configuration.

properties_file_content: Base64 encoded contents of connect-distributed.properties file.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A summary description of the worker configuration.

Main.Kafkaconnect.delete_connector — Method

delete_connector(connector_arn)
delete_connector(connector_arn, params::Dict{String,<:Any})

Deletes the specified connector.

Arguments

connector_arn: The Amazon Resource Name (ARN) of the connector that you want to delete.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"currentVersion": The current version of the connector that you want to delete.

Main.Kafkaconnect.describe_connector — Method

describe_connector(connector_arn)
describe_connector(connector_arn, params::Dict{String,<:Any})

Returns summary information about the connector.

Arguments

connector_arn: The Amazon Resource Name (ARN) of the connector that you want to describe.

Main.Kafkaconnect.describe_custom_plugin — Method

describe_custom_plugin(custom_plugin_arn)
describe_custom_plugin(custom_plugin_arn, params::Dict{String,<:Any})

A summary description of the custom plugin.

Arguments

custom_plugin_arn: Returns information about a custom plugin.

Main.Kafkaconnect.describe_worker_configuration — Method

describe_worker_configuration(worker_configuration_arn)
describe_worker_configuration(worker_configuration_arn, params::Dict{String,<:Any})

Returns information about a worker configuration.

Arguments

worker_configuration_arn: The Amazon Resource Name (ARN) of the worker configuration that you want to get information about.

Main.Kafkaconnect.list_connectors — Method

list_connectors()
list_connectors(params::Dict{String,<:Any})

Returns a list of all the connectors in this account and Region. The list is limited to connectors whose name starts with the specified prefix. The response also includes a description of each of the listed connectors.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"connectorNamePrefix": The name prefix that you want to use to search for and list connectors.

"maxResults": The maximum number of connectors to list in one response.

"nextToken": If the response of a ListConnectors operation is truncated, it will include a NextToken. Send this NextToken in a subsequent request to continue listing from where the previous operation left off.

Main.Kafkaconnect.list_custom_plugins — Method

list_custom_plugins()
list_custom_plugins(params::Dict{String,<:Any})

Returns a list of all of the custom plugins in this account and Region.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of custom plugins to list in one response.

"nextToken": If the response of a ListCustomPlugins operation is truncated, it will include a NextToken. Send this NextToken in a subsequent request to continue listing from where the previous operation left off.

Main.Kafkaconnect.list_worker_configurations — Method

list_worker_configurations()
list_worker_configurations(params::Dict{String,<:Any})

Returns a list of all of the worker configurations in this account and Region.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of worker configurations to list in one response.

"nextToken": If the response of a ListWorkerConfigurations operation is truncated, it will include a NextToken. Send this NextToken in a subsequent request to continue listing from where the previous operation left off.

Main.Kafkaconnect.update_connector — Method

update_connector(capacity, connector_arn, current_version)
update_connector(capacity, connector_arn, current_version, params::Dict{String,<:Any})

Updates the specified connector.

Arguments

capacity: The target capacity.

connector_arn: The Amazon Resource Name (ARN) of the connector that you want to update.

current_version: The current version of the connector that you want to update.