Kinesis Analytics V2

This page documents function available when using the Kinesis_Analytics_V2 module, created with @service Kinesis_Analytics_V2.

Index

Main.Kinesis_Analytics_V2.add_application_cloud_watch_logging_option

Main.Kinesis_Analytics_V2.add_application_input

Main.Kinesis_Analytics_V2.add_application_input_processing_configuration

Main.Kinesis_Analytics_V2.add_application_output

Main.Kinesis_Analytics_V2.add_application_reference_data_source

Main.Kinesis_Analytics_V2.add_application_vpc_configuration

Main.Kinesis_Analytics_V2.create_application

Main.Kinesis_Analytics_V2.create_application_presigned_url

Main.Kinesis_Analytics_V2.create_application_snapshot

Main.Kinesis_Analytics_V2.delete_application

Main.Kinesis_Analytics_V2.delete_application_cloud_watch_logging_option

Main.Kinesis_Analytics_V2.delete_application_input_processing_configuration

Main.Kinesis_Analytics_V2.delete_application_output

Main.Kinesis_Analytics_V2.delete_application_reference_data_source

Main.Kinesis_Analytics_V2.delete_application_snapshot

Main.Kinesis_Analytics_V2.delete_application_vpc_configuration

Main.Kinesis_Analytics_V2.describe_application

Main.Kinesis_Analytics_V2.describe_application_snapshot

Main.Kinesis_Analytics_V2.describe_application_version

Main.Kinesis_Analytics_V2.discover_input_schema

Main.Kinesis_Analytics_V2.list_application_snapshots

Main.Kinesis_Analytics_V2.list_application_versions

Main.Kinesis_Analytics_V2.list_applications

Main.Kinesis_Analytics_V2.list_tags_for_resource

Main.Kinesis_Analytics_V2.rollback_application

Main.Kinesis_Analytics_V2.start_application

Main.Kinesis_Analytics_V2.stop_application

Main.Kinesis_Analytics_V2.tag_resource

Main.Kinesis_Analytics_V2.untag_resource

Main.Kinesis_Analytics_V2.update_application

Main.Kinesis_Analytics_V2.update_application_maintenance_configuration

Documentation

Main.Kinesis_Analytics_V2.add_application_cloud_watch_logging_option — Method

add_application_cloud_watch_logging_option(application_name, cloud_watch_logging_option)
add_application_cloud_watch_logging_option(application_name, cloud_watch_logging_option, params::Dict{String,<:Any})

Adds an Amazon CloudWatch log stream to monitor application configuration errors.

Arguments

application_name: The Kinesis Data Analytics application name.

cloud_watch_logging_option: Provides the Amazon CloudWatch log stream Amazon Resource Name (ARN).

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ConditionalToken": A value you use to implement strong concurrency for application updates. You must provide the CurrentApplicationVersionId or the ConditionalToken. You get the application's current ConditionalToken using DescribeApplication. For better concurrency support, use the ConditionalToken parameter instead of CurrentApplicationVersionId.

"CurrentApplicationVersionId": The version ID of the Kinesis Data Analytics application. You must provide the CurrentApplicationVersionId or the ConditionalToken.You can retrieve the application version ID using DescribeApplication. For better concurrency support, use the ConditionalToken parameter instead of CurrentApplicationVersionId.

Main.Kinesis_Analytics_V2.add_application_input — Method

add_application_input(application_name, current_application_version_id, input)
add_application_input(application_name, current_application_version_id, input, params::Dict{String,<:Any})

Adds a streaming source to your SQL-based Kinesis Data Analytics application. You can add a streaming source when you create an application, or you can use this operation to add a streaming source after you create an application. For more information, see CreateApplication. Any configuration update, including adding a streaming source using this operation, results in a new version of the application. You can use the DescribeApplication operation to find the current application version.

Arguments

application_name: The name of your existing application to which you want to add the streaming source.

current_application_version_id: The current version of your application. You must provide the ApplicationVersionID or the ConditionalToken.You can use the DescribeApplication operation to find the current application version.

input: The Input to add.

Main.Kinesis_Analytics_V2.add_application_input_processing_configuration — Method

add_application_input_processing_configuration(application_name, current_application_version_id, input_id, input_processing_configuration)
add_application_input_processing_configuration(application_name, current_application_version_id, input_id, input_processing_configuration, params::Dict{String,<:Any})

Adds an InputProcessingConfiguration to a SQL-based Kinesis Data Analytics application. An input processor pre-processes records on the input stream before the application's SQL code executes. Currently, the only input processor available is Amazon Lambda.

Arguments

application_name: The name of the application to which you want to add the input processing configuration.

current_application_version_id: The version of the application to which you want to add the input processing configuration. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

input_id: The ID of the input configuration to add the input processing configuration to. You can get a list of the input IDs for an application using the DescribeApplication operation.

input_processing_configuration: The InputProcessingConfiguration to add to the application.

Main.Kinesis_Analytics_V2.add_application_output — Method

add_application_output(application_name, current_application_version_id, output)
add_application_output(application_name, current_application_version_id, output, params::Dict{String,<:Any})

Adds an external destination to your SQL-based Kinesis Data Analytics application. If you want Kinesis Data Analytics to deliver data from an in-application stream within your application to an external destination (such as an Kinesis data stream, a Kinesis Data Firehose delivery stream, or an Amazon Lambda function), you add the relevant configuration to your application using this operation. You can configure one or more outputs for your application. Each output configuration maps an in-application stream and an external destination. You can use one of the output configurations to deliver data from your in-application error stream to an external destination so that you can analyze the errors. Any configuration update, including adding a streaming source using this operation, results in a new version of the application. You can use the DescribeApplication operation to find the current application version.

Arguments

application_name: The name of the application to which you want to add the output configuration.

current_application_version_id: The version of the application to which you want to add the output configuration. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

output: An array of objects, each describing one output configuration. In the output configuration, you specify the name of an in-application stream, a destination (that is, a Kinesis data stream, a Kinesis Data Firehose delivery stream, or an Amazon Lambda function), and record the formation to use when writing to the destination.

Main.Kinesis_Analytics_V2.add_application_reference_data_source — Method

add_application_reference_data_source(application_name, current_application_version_id, reference_data_source)
add_application_reference_data_source(application_name, current_application_version_id, reference_data_source, params::Dict{String,<:Any})

Adds a reference data source to an existing SQL-based Kinesis Data Analytics application. Kinesis Data Analytics reads reference data (that is, an Amazon S3 object) and creates an in-application table within your application. In the request, you provide the source (S3 bucket name and object key name), name of the in-application table to create, and the necessary mapping information that describes how data in an Amazon S3 object maps to columns in the resulting in-application table.

Arguments

application_name: The name of an existing application.

current_application_version_id: The version of the application for which you are adding the reference data source. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

reference_data_source: The reference data source can be an object in your Amazon S3 bucket. Kinesis Data Analytics reads the object and copies the data into the in-application table that is created. You provide an S3 bucket, object key name, and the resulting in-application table that is created.

Main.Kinesis_Analytics_V2.add_application_vpc_configuration — Method

add_application_vpc_configuration(application_name, vpc_configuration)
add_application_vpc_configuration(application_name, vpc_configuration, params::Dict{String,<:Any})

Adds a Virtual Private Cloud (VPC) configuration to the application. Applications can use VPCs to store and access resources securely. Note the following about VPC configurations for Kinesis Data Analytics applications: VPC configurations are not supported for SQL applications. When a VPC is added to a Kinesis Data Analytics application, the application can no longer be accessed from the Internet directly. To enable Internet access to the application, add an Internet gateway to your VPC.

Arguments

application_name: The name of an existing application.

vpc_configuration: Description of the VPC to add to the application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ConditionalToken": A value you use to implement strong concurrency for application updates. You must provide the ApplicationVersionID or the ConditionalToken. You get the application's current ConditionalToken using DescribeApplication. For better concurrency support, use the ConditionalToken parameter instead of CurrentApplicationVersionId.

"CurrentApplicationVersionId": The version of the application to which you want to add the VPC configuration. You must provide the CurrentApplicationVersionId or the ConditionalToken. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned. For better concurrency support, use the ConditionalToken parameter instead of CurrentApplicationVersionId.

Main.Kinesis_Analytics_V2.create_application — Method

create_application(application_name, runtime_environment, service_execution_role)
create_application(application_name, runtime_environment, service_execution_role, params::Dict{String,<:Any})

Creates a Kinesis Data Analytics application. For information about creating a Kinesis Data Analytics application, see Creating an Application.

Arguments

application_name: The name of your application (for example, sample-app).

runtime_environment: The runtime environment for the application.

service_execution_role: The IAM role used by the application to access Kinesis data streams, Kinesis Data Firehose delivery streams, Amazon S3 objects, and other external resources.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ApplicationConfiguration": Use this parameter to configure the application.

"ApplicationDescription": A summary description of the application.

"ApplicationMode": Use the STREAMING mode to create a Kinesis Data Analytics For Flink application. To create a Kinesis Data Analytics Studio notebook, use the INTERACTIVE mode.

"CloudWatchLoggingOptions": Use this parameter to configure an Amazon CloudWatch log stream to monitor application configuration errors.

"Tags": A list of one or more tags to assign to the application. A tag is a key-value pair that identifies an application. Note that the maximum number of application tags includes system tags. The maximum number of user-defined application tags is 50. For more information, see Using Tagging.

Main.Kinesis_Analytics_V2.create_application_presigned_url — Method

create_application_presigned_url(application_name, url_type)
create_application_presigned_url(application_name, url_type, params::Dict{String,<:Any})

Creates and returns a URL that you can use to connect to an application's extension. The IAM role or user used to call this API defines the permissions to access the extension. After the presigned URL is created, no additional permission is required to access this URL. IAM authorization policies for this API are also enforced for every HTTP request that attempts to connect to the extension. You control the amount of time that the URL will be valid using the SessionExpirationDurationInSeconds parameter. If you do not provide this parameter, the returned URL is valid for twelve hours. The URL that you get from a call to CreateApplicationPresignedUrl must be used within 3 minutes to be valid. If you first try to use the URL after the 3-minute limit expires, the service returns an HTTP 403 Forbidden error.

Arguments

application_name: The name of the application.

url_type: The type of the extension for which to create and return a URL. Currently, the only valid extension URL type is FLINKDASHBOARDURL.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"SessionExpirationDurationInSeconds": The duration in seconds for which the returned URL will be valid.

Main.Kinesis_Analytics_V2.create_application_snapshot — Method

create_application_snapshot(application_name, snapshot_name)
create_application_snapshot(application_name, snapshot_name, params::Dict{String,<:Any})

Creates a snapshot of the application's state data.

Arguments

application_name: The name of an existing application

snapshot_name: An identifier for the application snapshot.

Main.Kinesis_Analytics_V2.delete_application — Method

delete_application(application_name, create_timestamp)
delete_application(application_name, create_timestamp, params::Dict{String,<:Any})

Deletes the specified application. Kinesis Data Analytics halts application execution and deletes the application.

Arguments

application_name: The name of the application to delete.

create_timestamp: Use the DescribeApplication operation to get this value.

Main.Kinesis_Analytics_V2.delete_application_cloud_watch_logging_option — Method

delete_application_cloud_watch_logging_option(application_name, cloud_watch_logging_option_id)
delete_application_cloud_watch_logging_option(application_name, cloud_watch_logging_option_id, params::Dict{String,<:Any})

Deletes an Amazon CloudWatch log stream from an Kinesis Data Analytics application.

Arguments

application_name: The application name.

cloud_watch_logging_option_id: The CloudWatchLoggingOptionId of the Amazon CloudWatch logging option to delete. You can get the CloudWatchLoggingOptionId by using the DescribeApplication operation.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ConditionalToken": A value you use to implement strong concurrency for application updates. You must provide the CurrentApplicationVersionId or the ConditionalToken. You get the application's current ConditionalToken using DescribeApplication. For better concurrency support, use the ConditionalToken parameter instead of CurrentApplicationVersionId.

"CurrentApplicationVersionId": The version ID of the application. You must provide the CurrentApplicationVersionId or the ConditionalToken. You can retrieve the application version ID using DescribeApplication. For better concurrency support, use the ConditionalToken parameter instead of CurrentApplicationVersionId.

Main.Kinesis_Analytics_V2.delete_application_input_processing_configuration — Method

delete_application_input_processing_configuration(application_name, current_application_version_id, input_id)
delete_application_input_processing_configuration(application_name, current_application_version_id, input_id, params::Dict{String,<:Any})

Deletes an InputProcessingConfiguration from an input.

Arguments

application_name: The name of the application.

current_application_version_id: The application version. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

input_id: The ID of the input configuration from which to delete the input processing configuration. You can get a list of the input IDs for an application by using the DescribeApplication operation.

Main.Kinesis_Analytics_V2.delete_application_output — Method

delete_application_output(application_name, current_application_version_id, output_id)
delete_application_output(application_name, current_application_version_id, output_id, params::Dict{String,<:Any})

Deletes the output destination configuration from your SQL-based Kinesis Data Analytics application's configuration. Kinesis Data Analytics will no longer write data from the corresponding in-application stream to the external output destination.

Arguments

application_name: The application name.

current_application_version_id: The application version. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

output_id: The ID of the configuration to delete. Each output configuration that is added to the application (either when the application is created or later) using the AddApplicationOutput operation has a unique ID. You need to provide the ID to uniquely identify the output configuration that you want to delete from the application configuration. You can use the DescribeApplication operation to get the specific OutputId.

Main.Kinesis_Analytics_V2.delete_application_reference_data_source — Method

delete_application_reference_data_source(application_name, current_application_version_id, reference_id)
delete_application_reference_data_source(application_name, current_application_version_id, reference_id, params::Dict{String,<:Any})

Deletes a reference data source configuration from the specified SQL-based Kinesis Data Analytics application's configuration. If the application is running, Kinesis Data Analytics immediately removes the in-application table that you created using the AddApplicationReferenceDataSource operation.

Arguments

application_name: The name of an existing application.

current_application_version_id: The current application version. You can use the DescribeApplication operation to get the current application version. If the version specified is not the current version, the ConcurrentModificationException is returned.

reference_id: The ID of the reference data source. When you add a reference data source to your application using the AddApplicationReferenceDataSource, Kinesis Data Analytics assigns an ID. You can use the DescribeApplication operation to get the reference ID.

Main.Kinesis_Analytics_V2.delete_application_snapshot — Method

delete_application_snapshot(application_name, snapshot_creation_timestamp, snapshot_name)
delete_application_snapshot(application_name, snapshot_creation_timestamp, snapshot_name, params::Dict{String,<:Any})

Deletes a snapshot of application state.

Arguments

application_name: The name of an existing application.

snapshot_creation_timestamp: The creation timestamp of the application snapshot to delete. You can retrieve this value using or .

snapshot_name: The identifier for the snapshot delete.

Main.Kinesis_Analytics_V2.delete_application_vpc_configuration — Method

delete_application_vpc_configuration(application_name, vpc_configuration_id)
delete_application_vpc_configuration(application_name, vpc_configuration_id, params::Dict{String,<:Any})

Removes a VPC configuration from a Kinesis Data Analytics application.

Arguments

application_name: The name of an existing application.

vpc_configuration_id: The ID of the VPC configuration to delete.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ConditionalToken": A value you use to implement strong concurrency for application updates. You must provide the CurrentApplicationVersionId or the ConditionalToken. You get the application's current ConditionalToken using DescribeApplication. For better concurrency support, use the ConditionalToken parameter instead of CurrentApplicationVersionId.

"CurrentApplicationVersionId": The current application version ID. You must provide the CurrentApplicationVersionId or the ConditionalToken. You can retrieve the application version ID using DescribeApplication. For better concurrency support, use the ConditionalToken parameter instead of CurrentApplicationVersionId.

Main.Kinesis_Analytics_V2.describe_application — Method

describe_application(application_name)
describe_application(application_name, params::Dict{String,<:Any})

Returns information about a specific Kinesis Data Analytics application. If you want to retrieve a list of all applications in your account, use the ListApplications operation.

Arguments

application_name: The name of the application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"IncludeAdditionalDetails": Displays verbose information about a Kinesis Data Analytics application, including the application's job plan.

Main.Kinesis_Analytics_V2.describe_application_snapshot — Method

describe_application_snapshot(application_name, snapshot_name)
describe_application_snapshot(application_name, snapshot_name, params::Dict{String,<:Any})

Returns information about a snapshot of application state data.

Arguments

application_name: The name of an existing application.

snapshot_name: The identifier of an application snapshot. You can retrieve this value using .

Main.Kinesis_Analytics_V2.describe_application_version — Method

describe_application_version(application_name, application_version_id)
describe_application_version(application_name, application_version_id, params::Dict{String,<:Any})

Provides a detailed description of a specified version of the application. To see a list of all the versions of an application, invoke the ListApplicationVersions operation. This operation is supported only for Amazon Kinesis Data Analytics for Apache Flink.

Arguments

application_name: The name of the application for which you want to get the version description.

application_version_id: The ID of the application version for which you want to get the description.

Main.Kinesis_Analytics_V2.discover_input_schema — Method

discover_input_schema(service_execution_role)
discover_input_schema(service_execution_role, params::Dict{String,<:Any})

Infers a schema for a SQL-based Kinesis Data Analytics application by evaluating sample records on the specified streaming source (Kinesis data stream or Kinesis Data Firehose delivery stream) or Amazon S3 object. In the response, the operation returns the inferred schema and also the sample records that the operation used to infer the schema. You can use the inferred schema when configuring a streaming source for your application. When you create an application using the Kinesis Data Analytics console, the console uses this operation to infer a schema and show it in the console user interface.

Arguments

service_execution_role: The ARN of the role that is used to access the streaming source.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"InputProcessingConfiguration": The InputProcessingConfiguration to use to preprocess the records before discovering the schema of the records.

"InputStartingPositionConfiguration": The point at which you want Kinesis Data Analytics to start reading records from the specified streaming source discovery purposes.

"ResourceARN": The Amazon Resource Name (ARN) of the streaming source.

"S3Configuration": Specify this parameter to discover a schema from data in an Amazon S3 object.

Main.Kinesis_Analytics_V2.list_application_snapshots — Method

list_application_snapshots(application_name)
list_application_snapshots(application_name, params::Dict{String,<:Any})

Lists information about the current application snapshots.

Arguments

application_name: The name of an existing application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Limit": The maximum number of application snapshots to list.

"NextToken": Use this parameter if you receive a NextToken response in a previous request that indicates that there is more output available. Set it to the value of the previous call's NextToken response to indicate where the output should continue from.

Main.Kinesis_Analytics_V2.list_application_versions — Method

list_application_versions(application_name)
list_application_versions(application_name, params::Dict{String,<:Any})

Lists all the versions for the specified application, including versions that were rolled back. The response also includes a summary of the configuration associated with each version. To get the complete description of a specific application version, invoke the DescribeApplicationVersion operation. This operation is supported only for Amazon Kinesis Data Analytics for Apache Flink.

Arguments

application_name: The name of the application for which you want to list all versions.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Limit": The maximum number of versions to list in this invocation of the operation.

"NextToken": If a previous invocation of this operation returned a pagination token, pass it into this value to retrieve the next set of results. For more information about pagination, see Using the Amazon Command Line Interface's Pagination Options.

Main.Kinesis_Analytics_V2.list_applications — Method

list_applications()
list_applications(params::Dict{String,<:Any})

Returns a list of Kinesis Data Analytics applications in your account. For each application, the response includes the application name, Amazon Resource Name (ARN), and status. If you want detailed information about a specific application, use DescribeApplication.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Limit": The maximum number of applications to list.

"NextToken": If a previous command returned a pagination token, pass it into this value to retrieve the next set of results. For more information about pagination, see Using the Amazon Command Line Interface's Pagination Options.

Main.Kinesis_Analytics_V2.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Retrieves the list of key-value tags assigned to the application. For more information, see Using Tagging.

Arguments

resource_arn: The ARN of the application for which to retrieve tags.

Main.Kinesis_Analytics_V2.rollback_application — Method

rollback_application(application_name, current_application_version_id)
rollback_application(application_name, current_application_version_id, params::Dict{String,<:Any})

Reverts the application to the previous running version. You can roll back an application if you suspect it is stuck in a transient status. You can roll back an application only if it is in the UPDATING or AUTOSCALING status. When you rollback an application, it loads state data from the last successful snapshot. If the application has no snapshots, Kinesis Data Analytics rejects the rollback request. This action is not supported for Kinesis Data Analytics for SQL applications.

Arguments

application_name: The name of the application.

current_application_version_id: The current application version ID. You can retrieve the application version ID using DescribeApplication.

Main.Kinesis_Analytics_V2.start_application — Method

start_application(application_name)
start_application(application_name, params::Dict{String,<:Any})

Starts the specified Kinesis Data Analytics application. After creating an application, you must exclusively call this operation to start your application.

Arguments

application_name: The name of the application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"RunConfiguration": Identifies the run configuration (start parameters) of a Kinesis Data Analytics application.

Main.Kinesis_Analytics_V2.stop_application — Method

stop_application(application_name)
stop_application(application_name, params::Dict{String,<:Any})

Stops the application from processing data. You can stop an application only if it is in the running status, unless you set the Force parameter to true. You can use the DescribeApplication operation to find the application status. Kinesis Data Analytics takes a snapshot when the application is stopped, unless Force is set to true.

Arguments

application_name: The name of the running application to stop.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Force": Set to true to force the application to stop. If you set Force to true, Kinesis Data Analytics stops the application without taking a snapshot. Force-stopping your application may lead to data loss or duplication. To prevent data loss or duplicate processing of data during application restarts, we recommend you to take frequent snapshots of your application. You can only force stop a Flink-based Kinesis Data Analytics application. You can't force stop a SQL-based Kinesis Data Analytics application. The application must be in the STARTING, UPDATING, STOPPING, AUTOSCALING, or RUNNING status.

Main.Kinesis_Analytics_V2.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Adds one or more key-value tags to a Kinesis Data Analytics application. Note that the maximum number of application tags includes system tags. The maximum number of user-defined application tags is 50. For more information, see Using Tagging.

Arguments

resource_arn: The ARN of the application to assign the tags.

tags: The key-value tags to assign to the application.

Main.Kinesis_Analytics_V2.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes one or more tags from a Kinesis Data Analytics application. For more information, see Using Tagging.

Arguments

resource_arn: The ARN of the Kinesis Data Analytics application from which to remove the tags.

tag_keys: A list of keys of tags to remove from the specified application.

Main.Kinesis_Analytics_V2.update_application — Method

update_application(application_name)
update_application(application_name, params::Dict{String,<:Any})

Updates an existing Kinesis Data Analytics application. Using this operation, you can update application code, input configuration, and output configuration. Kinesis Data Analytics updates the ApplicationVersionId each time you update your application. You cannot update the RuntimeEnvironment of an existing application. If you need to update an application's RuntimeEnvironment, you must delete the application and create it again.

Arguments

application_name: The name of the application to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ApplicationConfigurationUpdate": Describes application configuration updates.

"CloudWatchLoggingOptionUpdates": Describes application Amazon CloudWatch logging option updates. You can only update existing CloudWatch logging options with this action. To add a new CloudWatch logging option, use AddApplicationCloudWatchLoggingOption.

"ConditionalToken": A value you use to implement strong concurrency for application updates. You must provide the CurrentApplicationVersionId or the ConditionalToken. You get the application's current ConditionalToken using DescribeApplication. For better concurrency support, use the ConditionalToken parameter instead of CurrentApplicationVersionId.

"CurrentApplicationVersionId": The current application version ID. You must provide the CurrentApplicationVersionId or the ConditionalToken.You can retrieve the application version ID using DescribeApplication. For better concurrency support, use the ConditionalToken parameter instead of CurrentApplicationVersionId.

"RunConfigurationUpdate": Describes updates to the application's starting parameters.

"ServiceExecutionRoleUpdate": Describes updates to the service execution role.

Main.Kinesis_Analytics_V2.update_application_maintenance_configuration — Method

update_application_maintenance_configuration(application_maintenance_configuration_update, application_name)
update_application_maintenance_configuration(application_maintenance_configuration_update, application_name, params::Dict{String,<:Any})

Updates the maintenance configuration of the Kinesis Data Analytics application. You can invoke this operation on an application that is in one of the two following states: READY or RUNNING. If you invoke it when the application is in a state other than these two states, it throws a ResourceInUseException. The service makes use of the updated configuration the next time it schedules maintenance for the application. If you invoke this operation after the service schedules maintenance, the service will apply the configuration update the next time it schedules maintenance for the application. This means that you might not see the maintenance configuration update applied to the maintenance process that follows a successful invocation of this operation, but to the following maintenance process instead. To see the current maintenance configuration of your application, invoke the DescribeApplication operation. For information about application maintenance, see Kinesis Data Analytics for Apache Flink Maintenance. This operation is supported only for Amazon Kinesis Data Analytics for Apache Flink.

Arguments

application_maintenance_configuration_update: Describes the application maintenance configuration update.

application_name: The name of the application for which you want to update the maintenance configuration.