License Manager User Subscriptions

This page documents function available when using the License_Manager_User_Subscriptions module, created with @service License_Manager_User_Subscriptions.

Index

Main.License_Manager_User_Subscriptions.associate_user

Main.License_Manager_User_Subscriptions.create_license_server_endpoint

Main.License_Manager_User_Subscriptions.delete_license_server_endpoint

Main.License_Manager_User_Subscriptions.deregister_identity_provider

Main.License_Manager_User_Subscriptions.disassociate_user

Main.License_Manager_User_Subscriptions.list_identity_providers

Main.License_Manager_User_Subscriptions.list_instances

Main.License_Manager_User_Subscriptions.list_license_server_endpoints

Main.License_Manager_User_Subscriptions.list_product_subscriptions

Main.License_Manager_User_Subscriptions.list_tags_for_resource

Main.License_Manager_User_Subscriptions.list_user_associations

Main.License_Manager_User_Subscriptions.register_identity_provider

Main.License_Manager_User_Subscriptions.start_product_subscription

Main.License_Manager_User_Subscriptions.stop_product_subscription

Main.License_Manager_User_Subscriptions.tag_resource

Main.License_Manager_User_Subscriptions.untag_resource

Main.License_Manager_User_Subscriptions.update_identity_provider_settings

Documentation

Main.License_Manager_User_Subscriptions.associate_user — Function

associate_user(identity_provider, instance_id, username)
associate_user(identity_provider, instance_id, username, params::Dict{String,<:Any})

Associates the user to an EC2 instance to utilize user-based subscriptions.

Note

Your estimated bill for charges on the number of users and related costs will take 48 hours to appear for billing periods that haven't closed (marked as Pending billing status) in Amazon Web Services Billing. For more information, see Viewing your monthly charges in the Amazon Web Services Billing User Guide.

Arguments

identity_provider: The identity provider for the user.

instance_id: The ID of the EC2 instance that provides the user-based subscription.

username: The user name from the identity provider.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Domain": The domain name of the Active Directory that contains information for the user to associate.

"Tags": The tags that apply for the user association.

Main.License_Manager_User_Subscriptions.create_license_server_endpoint — Function

create_license_server_endpoint(identity_provider_arn, license_server_settings)
create_license_server_endpoint(identity_provider_arn, license_server_settings, params::Dict{String,<:Any})

Creates a network endpoint for the Remote Desktop Services (RDS) license server.

Arguments

identity_provider_arn: The Amazon Resource Name (ARN) that identifies the IdentityProvider resource that contains details about a registered identity provider. In the case of Active Directory, that can be a self-managed Active Directory or an Amazon Web Services Managed Active Directory that contains user identity details.

license_server_settings: The LicenseServerSettings resource to create for the endpoint. The settings include the type of license server and the Secrets Manager secret that enables administrators to add or remove users associated with the license server.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Tags": The tags that apply for the license server endpoint.

Main.License_Manager_User_Subscriptions.delete_license_server_endpoint — Function

delete_license_server_endpoint(license_server_endpoint_arn, server_type)
delete_license_server_endpoint(license_server_endpoint_arn, server_type, params::Dict{String,<:Any})

Deletes a LicenseServerEndpoint resource.

Arguments

license_server_endpoint_arn: The Amazon Resource Name (ARN) that identifies the LicenseServerEndpoint resource to delete.

server_type: The type of License Server that the delete request refers to.

Main.License_Manager_User_Subscriptions.deregister_identity_provider — Function

deregister_identity_provider()
deregister_identity_provider(params::Dict{String,<:Any})

Deregisters the Active Directory identity provider from License Manager user-based subscriptions.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"IdentityProvider": An object that specifies details for the Active Directory identity provider.

"IdentityProviderArn": The Amazon Resource Name (ARN) that identifies the identity provider to deregister.

"Product": The name of the user-based subscription product.

Valid values: VISUAL_STUDIO_ENTERPRISE | VISUAL_STUDIO_PROFESSIONAL | OFFICE_PROFESSIONAL_PLUS | REMOTE_DESKTOP_SERVICES

Main.License_Manager_User_Subscriptions.disassociate_user — Function

disassociate_user()
disassociate_user(params::Dict{String,<:Any})

Disassociates the user from an EC2 instance providing user-based subscriptions.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Domain": The domain name of the Active Directory that contains information for the user to disassociate.

"IdentityProvider": An object that specifies details for the Active Directory identity provider.

"InstanceId": The ID of the EC2 instance which provides user-based subscriptions.

"InstanceUserArn": The Amazon Resource Name (ARN) of the user to disassociate from the EC2 instance.

"Username": The user name from the Active Directory identity provider for the user.

Main.License_Manager_User_Subscriptions.list_identity_providers — Function

list_identity_providers()
list_identity_providers(params::Dict{String,<:Any})

Lists the Active Directory identity providers for user-based subscriptions.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Filters": You can use the following filters to streamline results:

Product

DirectoryId

"MaxResults": The maximum number of results to return from a single request.

"NextToken": A token to specify where to start paginating. This is the nextToken from a previously truncated response.

Main.License_Manager_User_Subscriptions.list_instances — Function

list_instances()
list_instances(params::Dict{String,<:Any})

Lists the EC2 instances providing user-based subscriptions.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Filters": You can use the following filters to streamline results:

Status

InstanceId

"MaxResults": The maximum number of results to return from a single request.

"NextToken": A token to specify where to start paginating. This is the nextToken from a previously truncated response.

Main.License_Manager_User_Subscriptions.list_license_server_endpoints — Function

list_license_server_endpoints()
list_license_server_endpoints(params::Dict{String,<:Any})

List the Remote Desktop Services (RDS) License Server endpoints

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Filters": You can use the following filters to streamline results:

IdentityProviderArn

"MaxResults": The maximum number of results to return from a single request.

"NextToken": A token to specify where to start paginating. This is the nextToken from a previously truncated response.

Main.License_Manager_User_Subscriptions.list_product_subscriptions — Function

list_product_subscriptions(identity_provider)
list_product_subscriptions(identity_provider, params::Dict{String,<:Any})

Lists the user-based subscription products available from an identity provider.

Arguments

identity_provider: An object that specifies details for the identity provider.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Filters": You can use the following filters to streamline results:

Status

Username

Domain

"MaxResults": The maximum number of results to return from a single request.

"NextToken": A token to specify where to start paginating. This is the nextToken from a previously truncated response.

"Product": The name of the user-based subscription product.

Valid values: VISUAL_STUDIO_ENTERPRISE | VISUAL_STUDIO_PROFESSIONAL | OFFICE_PROFESSIONAL_PLUS | REMOTE_DESKTOP_SERVICES

Main.License_Manager_User_Subscriptions.list_tags_for_resource — Function

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Returns the list of tags for the specified resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource whose tags you want to retrieve.

Main.License_Manager_User_Subscriptions.list_user_associations — Function

list_user_associations(identity_provider, instance_id)
list_user_associations(identity_provider, instance_id, params::Dict{String,<:Any})

Lists user associations for an identity provider.

Arguments

identity_provider: An object that specifies details for the identity provider.

instance_id: The ID of the EC2 instance, which provides user-based subscriptions.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Filters": You can use the following filters to streamline results:

Status

Username

Domain

"MaxResults": The maximum number of results to return from a single request.

"NextToken": A token to specify where to start paginating. This is the nextToken from a previously truncated response.

Main.License_Manager_User_Subscriptions.register_identity_provider — Function

register_identity_provider(identity_provider, product)
register_identity_provider(identity_provider, product, params::Dict{String,<:Any})

Registers an identity provider for user-based subscriptions.

Arguments

identity_provider: An object that specifies details for the identity provider to register.

product: The name of the user-based subscription product.

Valid values: VISUAL_STUDIO_ENTERPRISE | VISUAL_STUDIO_PROFESSIONAL | OFFICE_PROFESSIONAL_PLUS | REMOTE_DESKTOP_SERVICES

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Settings": The registered identity provider’s product related configuration settings such as the subnets to provision VPC endpoints.

"Tags": The tags that apply to the identity provider's registration.

Main.License_Manager_User_Subscriptions.start_product_subscription — Function

start_product_subscription(identity_provider, product, username)
start_product_subscription(identity_provider, product, username, params::Dict{String,<:Any})

Starts a product subscription for a user with the specified identity provider.

Note

Your estimated bill for charges on the number of users and related costs will take 48 hours to appear for billing periods that haven't closed (marked as Pending billing status) in Amazon Web Services Billing. For more information, see Viewing your monthly charges in the Amazon Web Services Billing User Guide.

Arguments

identity_provider: An object that specifies details for the identity provider.

product: The name of the user-based subscription product.

Valid values: VISUAL_STUDIO_ENTERPRISE | VISUAL_STUDIO_PROFESSIONAL | OFFICE_PROFESSIONAL_PLUS | REMOTE_DESKTOP_SERVICES

username: The user name from the identity provider of the user.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Domain": The domain name of the Active Directory that contains the user for whom to start the product subscription.

"Tags": The tags that apply to the product subscription.

Main.License_Manager_User_Subscriptions.stop_product_subscription — Function

stop_product_subscription()
stop_product_subscription(params::Dict{String,<:Any})

Stops a product subscription for a user with the specified identity provider.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Domain": The domain name of the Active Directory that contains the user for whom to stop the product subscription.

"IdentityProvider": An object that specifies details for the identity provider.

"Product": The name of the user-based subscription product.

Valid values: VISUAL_STUDIO_ENTERPRISE | VISUAL_STUDIO_PROFESSIONAL | OFFICE_PROFESSIONAL_PLUS | REMOTE_DESKTOP_SERVICES

"ProductUserArn": The Amazon Resource Name (ARN) of the product user.

"Username": The user name from the identity provider for the user.

Main.License_Manager_User_Subscriptions.tag_resource — Function

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Adds tags to a resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource that you want to tag.

tags: The tags to apply to the specified resource.

Main.License_Manager_User_Subscriptions.untag_resource — Function

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes tags from a resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource that you want to remove tags from.

tag_keys: The tag keys to remove from the resource.

Main.License_Manager_User_Subscriptions.update_identity_provider_settings — Function

update_identity_provider_settings(update_settings)
update_identity_provider_settings(update_settings, params::Dict{String,<:Any})

Updates additional product configuration settings for the registered identity provider.

Arguments

update_settings: Updates the registered identity provider’s product related configuration settings. You can update any combination of settings in a single operation such as the:

Subnets which you want to add to provision VPC endpoints.

Subnets which you want to remove the VPC endpoints from.

Security group ID which permits traffic to the VPC endpoints.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"IdentityProvider":

"IdentityProviderArn": The Amazon Resource Name (ARN) of the identity provider to update.

"Product": The name of the user-based subscription product.

Valid values: VISUAL_STUDIO_ENTERPRISE | VISUAL_STUDIO_PROFESSIONAL | OFFICE_PROFESSIONAL_PLUS | REMOTE_DESKTOP_SERVICES