Location

This page documents function available when using the Location module, created with @service Location.

Index

Main.Location.associate_tracker_consumer

Main.Location.batch_delete_geofence

Main.Location.batch_evaluate_geofences

Main.Location.batch_get_device_position

Main.Location.batch_put_geofence

Main.Location.batch_update_device_position

Main.Location.create_geofence_collection

Main.Location.create_map

Main.Location.create_place_index

Main.Location.create_tracker

Main.Location.delete_geofence_collection

Main.Location.delete_map

Main.Location.delete_place_index

Main.Location.delete_tracker

Main.Location.describe_geofence_collection

Main.Location.describe_map

Main.Location.describe_place_index

Main.Location.describe_tracker

Main.Location.disassociate_tracker_consumer

Main.Location.get_device_position

Main.Location.get_device_position_history

Main.Location.get_geofence

Main.Location.get_map_glyphs

Main.Location.get_map_sprites

Main.Location.get_map_style_descriptor

Main.Location.get_map_tile

Main.Location.list_geofence_collections

Main.Location.list_geofences

Main.Location.list_maps

Main.Location.list_place_indexes

Main.Location.list_tracker_consumers

Main.Location.list_trackers

Main.Location.put_geofence

Main.Location.search_place_index_for_position

Main.Location.search_place_index_for_text

Documentation

Main.Location.associate_tracker_consumer — Method

associate_tracker_consumer(consumer_arn, tracker_name)
associate_tracker_consumer(consumer_arn, tracker_name, params::Dict{String,<:Any})

Creates an association between a geofence collection and a tracker resource. This allows the tracker resource to communicate location data to the linked geofence collection. Currently not supported — Cross-account configurations, such as creating associations between a tracker resource in one account and a geofence collection in another account.

Arguments

consumer_arn: The Amazon Resource Name (ARN) for the geofence collection to be associated to tracker resource. Used when you need to specify a resource across all AWS. Format example: arn:partition:service:region:account-id:resource-type:resource-id

tracker_name: The name of the tracker resource to be associated with a geofence collection.

Main.Location.batch_delete_geofence — Method

batch_delete_geofence(collection_name, geofence_ids)
batch_delete_geofence(collection_name, geofence_ids, params::Dict{String,<:Any})

Deletes a batch of geofences from a geofence collection. This action deletes the resource permanently. You can't undo this action.

Arguments

collection_name: The geofence collection storing the geofences to be deleted.

geofence_ids: The batch of geofences to be deleted.

Main.Location.batch_evaluate_geofences — Method

batch_evaluate_geofences(collection_name, device_position_updates)
batch_evaluate_geofences(collection_name, device_position_updates, params::Dict{String,<:Any})

Evaluates device positions against the geofence geometries from a given geofence collection. The evaluation determines if the device has entered or exited a geofenced area, which publishes ENTER or EXIT geofence events to Amazon EventBridge. The last geofence that a device was observed within, if any, is tracked for 30 days after the most recent device position update

Arguments

collection_name: The geofence collection used in evaluating the position of devices against its geofences.

device_position_updates: Contains device details for each device to be evaluated against the given geofence collection.

Main.Location.batch_get_device_position — Method

batch_get_device_position(device_ids, tracker_name)
batch_get_device_position(device_ids, tracker_name, params::Dict{String,<:Any})

A batch request to retrieve all device positions.

Arguments

device_ids: Devices whose position you want to retrieve. For example, for two devices: device-ids=DeviceId1&amp;device-ids=DeviceId2

tracker_name: The tracker resource retrieving the device position.

Main.Location.batch_put_geofence — Method

batch_put_geofence(collection_name, entries)
batch_put_geofence(collection_name, entries, params::Dict{String,<:Any})

A batch request for storing geofence geometries into a given geofence collection.

Arguments

collection_name: The geofence collection storing the geofences.

entries: The batch of geofences to be stored in a geofence collection.

Main.Location.batch_update_device_position — Method

batch_update_device_position(tracker_name, updates)
batch_update_device_position(tracker_name, updates, params::Dict{String,<:Any})

Uploads position update data for one or more devices to a tracker resource. Amazon Location uses the data when reporting the last known device position and position history. Only one position update is stored per sample time. Location data is sampled at a fixed rate of one position per 30-second interval, and retained for one year before it is deleted.

Arguments

tracker_name: The name of the tracker resource to update.

updates: Contains the position update details for each device.

Main.Location.create_geofence_collection — Method

create_geofence_collection(collection_name, pricing_plan)
create_geofence_collection(collection_name, pricing_plan, params::Dict{String,<:Any})

Creates a geofence collection, which manages and stores geofences.

Arguments

collection_name: A custom name for the geofence collection. Requirements: Contain only alphanumeric characters (A–Z, a–z, 0-9), hyphens (-), periods (.), and underscores (_). Must be a unique geofence collection name. No spaces allowed. For example, ExampleGeofenceCollection.

pricing_plan: Specifies the pricing plan for your geofence collection. For additional details and restrictions on each pricing plan option, see the Amazon Location Service pricing page.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": An optional description for the geofence collection.

"PricingPlanDataSource": Specifies the plan data source. Required if the Mobile Asset Tracking (MAT) or the Mobile Asset Management (MAM) pricing plan is selected. Billing is determined by the resource usage, the associated pricing plan, and the data source that was specified. For more information about each pricing plan option and restrictions, see the Amazon Location Service pricing page. Valid Values: Esri | Here

Main.Location.create_map — Method

create_map(configuration, map_name, pricing_plan)
create_map(configuration, map_name, pricing_plan, params::Dict{String,<:Any})

Creates a map resource in your AWS account, which provides map tiles of different styles sourced from global location data providers. By using Maps, you agree that AWS may transmit your API queries to your selected third party provider for processing, which may be outside the AWS region you are currently using. For more information, see the AWS Service Terms for Amazon Location Service.

Arguments

configuration: Specifies the map style selected from an available data provider.

map_name: The name for the map resource. Requirements: Must contain only alphanumeric characters (A–Z, a–z, 0–9), hyphens (-), periods (.), and underscores (_). Must be a unique map resource name. No spaces allowed. For example, ExampleMap.

pricing_plan: Specifies the pricing plan for your map resource. For additional details and restrictions on each pricing plan option, see the Amazon Location Service pricing page.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": An optional description for the map resource.

Main.Location.create_place_index — Method

create_place_index(data_source, index_name, pricing_plan)
create_place_index(data_source, index_name, pricing_plan, params::Dict{String,<:Any})

Creates a Place index resource in your AWS account, which supports Places functions with geospatial data sourced from your chosen data provider. By using Places, you agree that AWS may transmit your API queries to your selected third party provider for processing, which may be outside the AWS region you are currently using. Because of licensing limitations, you may not use HERE to store results for locations in Japan. For more information, see the AWS Service Terms for Amazon Location Service.

Arguments

data_source: Specifies the data provider of geospatial data. This field is case-sensitive. Enter the valid values as shown. For example, entering HERE will return an error. Valid values include: Esri Here For additional details on data providers, see the Amazon Location Service data providers page.

index_name: The name of the Place index resource. Requirements: Contain only alphanumeric characters (A-Z, a-z, 0-9) , hyphens (-), periods (.), and underscores (_). Must be a unique Place index resource name. No spaces allowed. For example, ExamplePlaceIndex.

pricing_plan: Specifies the pricing plan for your Place index resource. For additional details and restrictions on each pricing plan option, see the Amazon Location Service pricing page.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DataSourceConfiguration": Specifies the data storage option for requesting Places.

"Description": The optional description for the Place index resource.

Main.Location.create_tracker — Method

create_tracker(pricing_plan, tracker_name)
create_tracker(pricing_plan, tracker_name, params::Dict{String,<:Any})

Creates a tracker resource in your AWS account, which lets you retrieve current and historical location of devices.

Arguments

pricing_plan: Specifies the pricing plan for your tracker resource. For additional details and restrictions on each pricing plan option, see the Amazon Location Service pricing page.

tracker_name: The name for the tracker resource. Requirements: Contain only alphanumeric characters (A-Z, a-z, 0-9) , hyphens (-), periods (.), and underscores (_). Must be a unique tracker resource name. No spaces allowed. For example, ExampleTracker.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": An optional description for the tracker resource.

"PricingPlanDataSource": Specifies the plan data source. Required if the Mobile Asset Tracking (MAT) or the Mobile Asset Management (MAM) pricing plan is selected. Billing is determined by the resource usage, the associated pricing plan, and data source that was specified. For more information about each pricing plan option and restrictions, see the Amazon Location Service pricing page. Valid Values: Esri | Here

Main.Location.delete_geofence_collection — Method

delete_geofence_collection(collection_name)
delete_geofence_collection(collection_name, params::Dict{String,<:Any})

Deletes a geofence collection from your AWS account. This action deletes the resource permanently. You can't undo this action. If the geofence collection is the target of a tracker resource, the devices will no longer be monitored.

Arguments

collection_name: The name of the geofence collection to be deleted.

Main.Location.delete_map — Method

delete_map(map_name)
delete_map(map_name, params::Dict{String,<:Any})

Deletes a map resource from your AWS account. This action deletes the resource permanently. You cannot undo this action. If the map is being used in an application, the map may not render.

Arguments

map_name: The name of the map resource to be deleted.

Main.Location.delete_place_index — Method

delete_place_index(index_name)
delete_place_index(index_name, params::Dict{String,<:Any})

Deletes a Place index resource from your AWS account. This action deletes the resource permanently. You cannot undo this action.

Arguments

index_name: The name of the Place index resource to be deleted.

Main.Location.delete_tracker — Method

delete_tracker(tracker_name)
delete_tracker(tracker_name, params::Dict{String,<:Any})

Deletes a tracker resource from your AWS account. This action deletes the resource permanently. You can't undo this action. If the tracker resource is in use, you may encounter an error. Make sure that the target resource is not a dependency for your applications.

Arguments

tracker_name: The name of the tracker resource to be deleted.

Main.Location.describe_geofence_collection — Method

describe_geofence_collection(collection_name)
describe_geofence_collection(collection_name, params::Dict{String,<:Any})

Retrieves the geofence collection details.

Arguments

collection_name: The name of the geofence collection.

Main.Location.describe_map — Method

describe_map(map_name)
describe_map(map_name, params::Dict{String,<:Any})

Retrieves the map resource details.

Arguments

map_name: The name of the map resource.

Main.Location.describe_place_index — Method

describe_place_index(index_name)
describe_place_index(index_name, params::Dict{String,<:Any})

Retrieves the Place index resource details.

Arguments

index_name: The name of the Place index resource.

Main.Location.describe_tracker — Method

describe_tracker(tracker_name)
describe_tracker(tracker_name, params::Dict{String,<:Any})

Retrieves the tracker resource details.

Arguments

tracker_name: The name of the tracker resource.

Main.Location.disassociate_tracker_consumer — Method

disassociate_tracker_consumer(consumer_arn, tracker_name)
disassociate_tracker_consumer(consumer_arn, tracker_name, params::Dict{String,<:Any})

Removes the association between a tracker resource and a geofence collection. Once you unlink a tracker resource from a geofence collection, the tracker positions will no longer be automatically evaluated against geofences.

Arguments

consumer_arn: The Amazon Resource Name (ARN) for the geofence collection to be disassociated from the tracker resource. Used when you need to specify a resource across all AWS. Format example: arn:partition:service:region:account-id:resource-type:resource-id

tracker_name: The name of the tracker resource to be dissociated from the consumer.

Main.Location.get_device_position — Method

get_device_position(device_id, tracker_name)
get_device_position(device_id, tracker_name, params::Dict{String,<:Any})

Retrieves a device's most recent position according to its sample time. Device positions are deleted after one year.

Arguments

device_id: The device whose position you want to retrieve.

tracker_name: The tracker resource receiving the position update.

Main.Location.get_device_position_history — Method

get_device_position_history(device_id, tracker_name)
get_device_position_history(device_id, tracker_name, params::Dict{String,<:Any})

Retrieves the device position history from a tracker resource within a specified range of time. Device positions are deleted after 1 year.

Arguments

device_id: The device whose position history you want to retrieve.

tracker_name: The tracker resource receiving the request for the device position history.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"EndTimeExclusive": Specify the end time for the position history in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ. By default, the value will be the time that the request is made. Requirement: The time specified for EndTimeExclusive must be after the time for StartTimeInclusive.

"NextToken": The pagination token specifying which page of results to return in the response. If no token is provided, the default page is the first page. Default value: null

"StartTimeInclusive": Specify the start time for the position history in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ. By default, the value will be 24 hours prior to the time that the request is made. Requirement: The time specified for StartTimeInclusive must be before EndTimeExclusive.

Main.Location.get_geofence — Method

get_geofence(collection_name, geofence_id)
get_geofence(collection_name, geofence_id, params::Dict{String,<:Any})

Retrieves the geofence details from a geofence collection.

Arguments

collection_name: The geofence collection storing the target geofence.

geofence_id: The geofence you're retrieving details for.

Main.Location.get_map_glyphs — Method

get_map_glyphs(font_stack, font_unicode_range, map_name)
get_map_glyphs(font_stack, font_unicode_range, map_name, params::Dict{String,<:Any})

Retrieves glyphs used to display labels on a map.

Arguments

font_stack: A comma-separated list of fonts to load glyphs from in order of preference.. For example, Noto Sans, Arial Unicode.

font_unicode_range: A Unicode range of characters to download glyphs for. Each response will contain 256 characters. For example, 0-255 includes all characters from range U+0000 to 00FF. Must be aligned to multiples of 256.

map_name: The map resource associated with the glyph ﬁle.

Main.Location.get_map_sprites — Method

get_map_sprites(file_name, map_name)
get_map_sprites(file_name, map_name, params::Dict{String,<:Any})

Retrieves the sprite sheet corresponding to a map resource. The sprite sheet is a PNG image paired with a JSON document describing the offsets of individual icons that will be displayed on a rendered map.

Arguments

file_name: The name of the sprite ﬁle. Use the following ﬁle names for the sprite sheet: sprites.png sprites@2x.png for high pixel density displays For the JSON document contain image offsets. Use the following ﬁle names: sprites.json sprites@2x.json for high pixel density displays

map_name: The map resource associated with the sprite ﬁle.

Main.Location.get_map_style_descriptor — Method

get_map_style_descriptor(map_name)
get_map_style_descriptor(map_name, params::Dict{String,<:Any})

Retrieves the map style descriptor from a map resource. The style descriptor contains speciﬁcations on how features render on a map. For example, what data to display, what order to display the data in, and the style for the data. Style descriptors follow the Mapbox Style Specification.

Arguments

map_name: The map resource to retrieve the style descriptor from.

Main.Location.get_map_tile — Method

get_map_tile(map_name, x, y, z)
get_map_tile(map_name, x, y, z, params::Dict{String,<:Any})

Retrieves a vector data tile from the map resource. Map tiles are used by clients to render a map. They are addressed using a grid arrangement with an X coordinate, Y coordinate, and Z (zoom) level. The origin (0, 0) is the top left of the map. Increasing the zoom level by 1 doubles both the X and Y dimensions, so a tile containing data for the entire world at (0/0/0) will be split into 4 tiles at zoom 1 (1/0/0, 1/0/1, 1/1/0, 1/1/1).

Arguments

map_name: The map resource to retrieve the map tiles from.

x: The X axis value for the map tile.

y: The Y axis value for the map tile.

z: The zoom value for the map tile.

Main.Location.list_geofence_collections — Method

list_geofence_collections()
list_geofence_collections(params::Dict{String,<:Any})

Lists geofence collections in your AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": An optional limit for the number of resources returned in a single call. Default value: 100

"NextToken": The pagination token specifying which page of results to return in the response. If no token is provided, the default page is the first page. Default value: null

Main.Location.list_geofences — Method

list_geofences(collection_name)
list_geofences(collection_name, params::Dict{String,<:Any})

Lists geofences stored in a given geofence collection.

Arguments

collection_name: The name of the geofence collection storing the list of geofences.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"NextToken": The pagination token specifying which page of results to return in the response. If no token is provided, the default page is the first page. Default value: null

Main.Location.list_maps — Method

list_maps()
list_maps(params::Dict{String,<:Any})

Lists map resources in your AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": An optional limit for the number of resources returned in a single call. Default value: 100

"NextToken": The pagination token specifying which page of results to return in the response. If no token is provided, the default page is the first page. Default value: null

Main.Location.list_place_indexes — Method

list_place_indexes()
list_place_indexes(params::Dict{String,<:Any})

Lists Place index resources in your AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": An optional limit for the maximum number of results returned in a single call. Default value: 100

"NextToken": The pagination token specifying which page of results to return in the response. If no token is provided, the default page is the first page. Default value: null

Main.Location.list_tracker_consumers — Method

list_tracker_consumers(tracker_name)
list_tracker_consumers(tracker_name, params::Dict{String,<:Any})

Lists geofence collections currently associated to the given tracker resource.

Arguments

tracker_name: The tracker resource whose associated geofence collections you want to list.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": An optional limit for the number of resources returned in a single call. Default value: 100

"NextToken": The pagination token specifying which page of results to return in the response. If no token is provided, the default page is the first page. Default value: null

Main.Location.list_trackers — Method

list_trackers()
list_trackers(params::Dict{String,<:Any})

Lists tracker resources in your AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": An optional limit for the number of resources returned in a single call. Default value: 100

"NextToken": The pagination token specifying which page of results to return in the response. If no token is provided, the default page is the first page. Default value: null

Main.Location.put_geofence — Method

put_geofence(collection_name, geofence_id, geometry)
put_geofence(collection_name, geofence_id, geometry, params::Dict{String,<:Any})

Stores a geofence geometry in a given geofence collection, or updates the geometry of an existing geofence if a geofence ID is included in the request.

Arguments

collection_name: The geofence collection to store the geofence in.

geofence_id: An identifier for the geofence. For example, ExampleGeofence-1.

geometry: Contains the polygon details to specify the position of the geofence. Each geofence polygon can have a maximum of 1,000 vertices.

Main.Location.search_place_index_for_position — Method

search_place_index_for_position(index_name, position)
search_place_index_for_position(index_name, position, params::Dict{String,<:Any})

Reverse geocodes a given coordinate and returns a legible address. Allows you to search for Places or points of interest near a given position. By using Places, you agree that AWS may transmit your API queries to your selected third party provider for processing, which may be outside the AWS region you are currently using. Because of licensing limitations, you may not use HERE to store results for locations in Japan. For more information, see the AWS Service Terms for Amazon Location Service.

Arguments

index_name: The name of the Place index resource you want to use for the search.

position: Specifies a coordinate for the query defined by a longitude, and latitude. The first position is the X coordinate, or longitude. The second position is the Y coordinate, or latitude. For example, position=xLongitude&amp;position=yLatitude .

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": An optional paramer. The maximum number of results returned per request. Default value: 50

Main.Location.search_place_index_for_text — Method

search_place_index_for_text(index_name, text)
search_place_index_for_text(index_name, text, params::Dict{String,<:Any})

Geocodes free-form text, such as an address, name, city, or region to allow you to search for Places or points of interest. Includes the option to apply additional parameters to narrow your list of results. You can search for places near a given position using BiasPosition, or filter results within a bounding box using FilterBBox. Providing both parameters simultaneously returns an error. By using Places, you agree that AWS may transmit your API queries to your selected third party provider for processing, which may be outside the AWS region you are currently using. Also, when using HERE as your data provider, you may not (a) use HERE Places for Asset Management, or (b) select the Storage option for the IntendedUse parameter when requesting Places in Japan. For more information, see the AWS Service Terms for Amazon Location Service.

Arguments

index_name: The name of the Place index resource you want to use for the search.

text: The address, name, city, or region to be used in the search. In free-form text format. For example, 123 Any Street.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"BiasPosition": Searches for results closest to the given position. An optional parameter defined by longitude, and latitude. The first bias position is the X coordinate, or longitude. The second bias position is the Y coordinate, or latitude. For example, bias=xLongitude&amp;bias=yLatitude.

"FilterBBox": Filters the results by returning only Places within the provided bounding box. An optional parameter. The first 2 bbox parameters describe the lower southwest corner: The first bbox position is the X coordinate or longitude of the lower southwest corner. The second bbox position is the Y coordinate or latitude of the lower southwest corner. For example, bbox=xLongitudeSW&amp;bbox=yLatitudeSW. The next bbox parameters describe the upper northeast corner: The third bbox position is the X coordinate, or longitude of the upper northeast corner. The fourth bbox position is the Y coordinate, or longitude of the upper northeast corner. For example, bbox=xLongitudeNE&amp;bbox=yLatitudeNE

"FilterCountries": Limits the search to the given a list of countries/regions. An optional parameter. Use the ISO 3166 3-digit country code. For example, Australia uses three upper-case characters: AUS.

"MaxResults": An optional parameter. The maximum number of results returned per request. The default: 50