M2

This page documents function available when using the M2 module, created with @service M2.

Index

Main.M2.cancel_batch_job_execution

Main.M2.create_application

Main.M2.create_data_set_import_task

Main.M2.create_deployment

Main.M2.create_environment

Main.M2.delete_application

Main.M2.delete_application_from_environment

Main.M2.delete_environment

Main.M2.get_application

Main.M2.get_application_version

Main.M2.get_batch_job_execution

Main.M2.get_data_set_details

Main.M2.get_data_set_import_task

Main.M2.get_deployment

Main.M2.get_environment

Main.M2.get_signed_bluinsights_url

Main.M2.list_application_versions

Main.M2.list_applications

Main.M2.list_batch_job_definitions

Main.M2.list_batch_job_executions

Main.M2.list_batch_job_restart_points

Main.M2.list_data_set_import_history

Main.M2.list_data_sets

Main.M2.list_deployments

Main.M2.list_engine_versions

Main.M2.list_environments

Main.M2.list_tags_for_resource

Main.M2.start_application

Main.M2.start_batch_job

Main.M2.stop_application

Main.M2.tag_resource

Main.M2.untag_resource

Main.M2.update_application

Main.M2.update_environment

Documentation

Main.M2.cancel_batch_job_execution — Method

cancel_batch_job_execution(application_id, execution_id)
cancel_batch_job_execution(application_id, execution_id, params::Dict{String,<:Any})

Cancels the running of a specific batch job execution.

Arguments

application_id: The unique identifier of the application.

execution_id: The unique identifier of the batch job execution.

Main.M2.create_application — Method

create_application(definition, engine_type, name)
create_application(definition, engine_type, name, params::Dict{String,<:Any})

Creates a new application with given parameters. Requires an existing runtime environment and application definition file.

Arguments

definition: The application definition for this application. You can specify either inline JSON or an S3 bucket location.

engine_type: The type of the target platform for this application.

name: The unique identifier of the application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier the service generates to ensure the idempotency of the request to create an application. The service generates the clientToken when the API call is triggered. The token expires after one hour, so if you retry the API within this timeframe with the same clientToken, you will get the same response. The service also handles deleting the clientToken after it expires.

"description": The description of the application.

"kmsKeyId": The identifier of a customer managed key.

"roleArn": The Amazon Resource Name (ARN) that identifies a role that the application uses to access Amazon Web Services resources that are not part of the application or are in a different Amazon Web Services account.

"tags": A list of tags to apply to the application.

Main.M2.create_data_set_import_task — Method

create_data_set_import_task(application_id, import_config)
create_data_set_import_task(application_id, import_config, params::Dict{String,<:Any})

Starts a data set import task for a specific application.

Arguments

application_id: The unique identifier of the application for which you want to import data sets.

import_config: The data set import task configuration.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier you provide to ensure the idempotency of the request to create a data set import. The service generates the clientToken when the API call is triggered. The token expires after one hour, so if you retry the API within this timeframe with the same clientToken, you will get the same response. The service also handles deleting the clientToken after it expires.

Main.M2.create_deployment — Method

create_deployment(application_id, application_version, environment_id)
create_deployment(application_id, application_version, environment_id, params::Dict{String,<:Any})

Creates and starts a deployment to deploy an application into a runtime environment.

Arguments

application_id: The application identifier.

application_version: The version of the application to deploy.

environment_id: The identifier of the runtime environment where you want to deploy this application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier you provide to ensure the idempotency of the request to create a deployment. The service generates the clientToken when the API call is triggered. The token expires after one hour, so if you retry the API within this timeframe with the same clientToken, you will get the same response. The service also handles deleting the clientToken after it expires.

Main.M2.create_environment — Method

create_environment(engine_type, instance_type, name)
create_environment(engine_type, instance_type, name, params::Dict{String,<:Any})

Creates a runtime environment for a given runtime engine.

Arguments

engine_type: The engine type for the runtime environment.

instance_type: The type of instance for the runtime environment.

name: The name of the runtime environment. Must be unique within the account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier you provide to ensure the idempotency of the request to create an environment. The service generates the clientToken when the API call is triggered. The token expires after one hour, so if you retry the API within this timeframe with the same clientToken, you will get the same response. The service also handles deleting the clientToken after it expires.

"description": The description of the runtime environment.

"engineVersion": The version of the engine type for the runtime environment.

"highAvailabilityConfig": The details of a high availability configuration for this runtime environment.

"kmsKeyId": The identifier of a customer managed key.

"preferredMaintenanceWindow": Configures the maintenance window that you want for the runtime environment. The maintenance window must have the format ddd:hh24:mi-ddd:hh24:mi and must be less than 24 hours. The following two examples are valid maintenance windows: sun:23:45-mon:00:15 or sat:01:00-sat:03:00. If you do not provide a value, a random system-generated value will be assigned.

"publiclyAccessible": Specifies whether the runtime environment is publicly accessible.

"securityGroupIds": The list of security groups for the VPC associated with this runtime environment.

"storageConfigurations": Optional. The storage configurations for this runtime environment.

"subnetIds": The list of subnets associated with the VPC for this runtime environment.

"tags": The tags for the runtime environment.

Main.M2.delete_application — Method

delete_application(application_id)
delete_application(application_id, params::Dict{String,<:Any})

Deletes a specific application. You cannot delete a running application.

Arguments

application_id: The unique identifier of the application you want to delete.

Main.M2.delete_application_from_environment — Method

delete_application_from_environment(application_id, environment_id)
delete_application_from_environment(application_id, environment_id, params::Dict{String,<:Any})

Deletes a specific application from the specific runtime environment where it was previously deployed. You cannot delete a runtime environment using DeleteEnvironment if any application has ever been deployed to it. This API removes the association of the application with the runtime environment so you can delete the environment smoothly.

Arguments

application_id: The unique identifier of the application you want to delete.

environment_id: The unique identifier of the runtime environment where the application was previously deployed.

Main.M2.delete_environment — Method

delete_environment(environment_id)
delete_environment(environment_id, params::Dict{String,<:Any})

Deletes a specific runtime environment. The environment cannot contain deployed applications. If it does, you must delete those applications before you delete the environment.

Arguments

environment_id: The unique identifier of the runtime environment you want to delete.

Main.M2.get_application — Method

get_application(application_id)
get_application(application_id, params::Dict{String,<:Any})

Describes the details of a specific application.

Arguments

application_id: The identifier of the application.

Main.M2.get_application_version — Method

get_application_version(application_id, application_version)
get_application_version(application_id, application_version, params::Dict{String,<:Any})

Returns details about a specific version of a specific application.

Arguments

application_id: The unique identifier of the application.

application_version: The specific version of the application.

Main.M2.get_batch_job_execution — Method

get_batch_job_execution(application_id, execution_id)
get_batch_job_execution(application_id, execution_id, params::Dict{String,<:Any})

Gets the details of a specific batch job execution for a specific application.

Arguments

application_id: The identifier of the application.

execution_id: The unique identifier of the batch job execution.

Main.M2.get_data_set_details — Method

get_data_set_details(application_id, data_set_name)
get_data_set_details(application_id, data_set_name, params::Dict{String,<:Any})

Gets the details of a specific data set.

Arguments

application_id: The unique identifier of the application that this data set is associated with.

data_set_name: The name of the data set.

Main.M2.get_data_set_import_task — Method

get_data_set_import_task(application_id, task_id)
get_data_set_import_task(application_id, task_id, params::Dict{String,<:Any})

Gets the status of a data set import task initiated with the CreateDataSetImportTask operation.

Arguments

application_id: The application identifier.

task_id: The task identifier returned by the CreateDataSetImportTask operation.

Main.M2.get_deployment — Method

get_deployment(application_id, deployment_id)
get_deployment(application_id, deployment_id, params::Dict{String,<:Any})

Gets details of a specific deployment with a given deployment identifier.

Arguments

application_id: The unique identifier of the application.

deployment_id: The unique identifier for the deployment.

Main.M2.get_environment — Method

get_environment(environment_id)
get_environment(environment_id, params::Dict{String,<:Any})

Describes a specific runtime environment.

Arguments

environment_id: The unique identifier of the runtime environment.

Main.M2.get_signed_bluinsights_url — Method

get_signed_bluinsights_url()
get_signed_bluinsights_url(params::Dict{String,<:Any})

Gets a single sign-on URL that can be used to connect to AWS Blu Insights.

Main.M2.list_application_versions — Method

list_application_versions(application_id)
list_application_versions(application_id, params::Dict{String,<:Any})

Returns a list of the application versions for a specific application.

Arguments

application_id: The unique identifier of the application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of application versions to return.

"nextToken": A pagination token returned from a previous call to this operation. This specifies the next item to return. To return to the beginning of the list, exclude this parameter.

Main.M2.list_applications — Method

list_applications()
list_applications(params::Dict{String,<:Any})

Lists the applications associated with a specific Amazon Web Services account. You can provide the unique identifier of a specific runtime environment in a query parameter to see all applications associated with that environment.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"environmentId": The unique identifier of the runtime environment where the applications are deployed.

"maxResults": The maximum number of applications to return.

"names": The names of the applications.

"nextToken": A pagination token to control the number of applications displayed in the list.

Main.M2.list_batch_job_definitions — Method

list_batch_job_definitions(application_id)
list_batch_job_definitions(application_id, params::Dict{String,<:Any})

Lists all the available batch job definitions based on the batch job resources uploaded during the application creation. You can use the batch job definitions in the list to start a batch job.

Arguments

application_id: The identifier of the application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of batch job definitions to return.

"nextToken": A pagination token returned from a previous call to this operation. This specifies the next item to return. To return to the beginning of the list, exclude this parameter.

"prefix": If the batch job definition is a FileBatchJobDefinition, the prefix allows you to search on the file names of FileBatchJobDefinitions.

Main.M2.list_batch_job_executions — Method

list_batch_job_executions(application_id)
list_batch_job_executions(application_id, params::Dict{String,<:Any})

Lists historical, current, and scheduled batch job executions for a specific application.

Arguments

application_id: The unique identifier of the application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"executionIds": The unique identifier of each batch job execution.

"jobName": The name of each batch job execution.

"maxResults": The maximum number of batch job executions to return.

"nextToken": A pagination token to control the number of batch job executions displayed in the list.

"startedAfter": The time after which the batch job executions started.

"startedBefore": The time before the batch job executions started.

"status": The status of the batch job executions.

Main.M2.list_batch_job_restart_points — Method

list_batch_job_restart_points(application_id, execution_id)
list_batch_job_restart_points(application_id, execution_id, params::Dict{String,<:Any})

Lists all the job steps for JCL files to restart a batch job. This is only applicable for Micro Focus engine with versions 8.0.6 and above.

Arguments

application_id: The unique identifier of the application.

execution_id: The unique identifier of each batch job execution.

Main.M2.list_data_set_import_history — Method

list_data_set_import_history(application_id)
list_data_set_import_history(application_id, params::Dict{String,<:Any})

Lists the data set imports for the specified application.

Arguments

application_id: The unique identifier of the application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of objects to return.

"nextToken": A pagination token returned from a previous call to this operation. This specifies the next item to return. To return to the beginning of the list, exclude this parameter.

Main.M2.list_data_sets — Method

list_data_sets(application_id)
list_data_sets(application_id, params::Dict{String,<:Any})

Lists the data sets imported for a specific application. In Amazon Web Services Mainframe Modernization, data sets are associated with applications deployed on runtime environments. This is known as importing data sets. Currently, Amazon Web Services Mainframe Modernization can import data sets into catalogs using CreateDataSetImportTask.

Arguments

application_id: The unique identifier of the application for which you want to list the associated data sets.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of objects to return.

"nameFilter": Filter dataset name matching the specified pattern. Can use * and % as wild cards.

"nextToken": A pagination token returned from a previous call to this operation. This specifies the next item to return. To return to the beginning of the list, exclude this parameter.

"prefix": The prefix of the data set name, which you can use to filter the list of data sets.

Main.M2.list_deployments — Method

list_deployments(application_id)
list_deployments(application_id, params::Dict{String,<:Any})

Returns a list of all deployments of a specific application. A deployment is a combination of a specific application and a specific version of that application. Each deployment is mapped to a particular application version.

Arguments

application_id: The application identifier.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of objects to return.

"nextToken": A pagination token returned from a previous call to this operation. This specifies the next item to return. To return to the beginning of the list, exclude this parameter.

Main.M2.list_engine_versions — Method

list_engine_versions()
list_engine_versions(params::Dict{String,<:Any})

Lists the available engine versions.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"engineType": The type of target platform.

"maxResults": The maximum number of objects to return.

"nextToken": A pagination token returned from a previous call to this operation. This specifies the next item to return. To return to the beginning of the list, exclude this parameter.

Main.M2.list_environments — Method

list_environments()
list_environments(params::Dict{String,<:Any})

Lists the runtime environments.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"engineType": The engine type for the runtime environment.

"maxResults": The maximum number of runtime environments to return.

"names": The names of the runtime environments. Must be unique within the account.

"nextToken": A pagination token to control the number of runtime environments displayed in the list.

Main.M2.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists the tags for the specified resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

Main.M2.start_application — Method

start_application(application_id)
start_application(application_id, params::Dict{String,<:Any})

Starts an application that is currently stopped.

Arguments

application_id: The unique identifier of the application you want to start.

Main.M2.start_batch_job — Method

start_batch_job(application_id, batch_job_identifier)
start_batch_job(application_id, batch_job_identifier, params::Dict{String,<:Any})

Starts a batch job and returns the unique identifier of this execution of the batch job. The associated application must be running in order to start the batch job.

Arguments

application_id: The unique identifier of the application associated with this batch job.

batch_job_identifier: The unique identifier of the batch job.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"jobParams": The collection of batch job parameters. For details about limits for keys and values, see Coding variables in JCL.

Main.M2.stop_application — Method

stop_application(application_id)
stop_application(application_id, params::Dict{String,<:Any})

Stops a running application.

Arguments

application_id: The unique identifier of the application you want to stop.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"forceStop": Stopping an application process can take a long time. Setting this parameter to true lets you force stop the application so you don't need to wait until the process finishes to apply another action on the application. The default value is false.

Main.M2.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Adds one or more tags to the specified resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

tags: The tags to add to the resource.

Main.M2.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes one or more tags from the specified resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

tag_keys: The keys of the tags to remove.

Main.M2.update_application — Method

update_application(application_id, current_application_version)
update_application(application_id, current_application_version, params::Dict{String,<:Any})

Updates an application and creates a new version.

Arguments

application_id: The unique identifier of the application you want to update.

current_application_version: The current version of the application to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"definition": The application definition for this application. You can specify either inline JSON or an S3 bucket location.

"description": The description of the application to update.

Main.M2.update_environment — Method

update_environment(environment_id)
update_environment(environment_id, params::Dict{String,<:Any})

Updates the configuration details for a specific runtime environment.

Arguments

environment_id: The unique identifier of the runtime environment that you want to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"applyDuringMaintenanceWindow": Indicates whether to update the runtime environment during the maintenance window. The default is false. Currently, Amazon Web Services Mainframe Modernization accepts the engineVersion parameter only if applyDuringMaintenanceWindow is true. If any parameter other than engineVersion is provided in UpdateEnvironmentRequest, it will fail if applyDuringMaintenanceWindow is set to true.

"desiredCapacity": The desired capacity for the runtime environment to update. The minimum possible value is 0 and the maximum is 100.

"engineVersion": The version of the runtime engine for the runtime environment.

"forceUpdate": Forces the updates on the environment. This option is needed if the applications in the environment are not stopped or if there are ongoing application-related activities in the environment. If you use this option, be aware that it could lead to data corruption in the applications, and that you might need to perform repair and recovery procedures for the applications. This option is not needed if the attribute being updated is preferredMaintenanceWindow.

"instanceType": The instance type for the runtime environment to update.

"preferredMaintenanceWindow": Configures the maintenance window that you want for the runtime environment. The maintenance window must have the format ddd:hh24:mi-ddd:hh24:mi and must be less than 24 hours. The following two examples are valid maintenance windows: sun:23:45-mon:00:15 or sat:01:00-sat:03:00. If you do not provide a value, a random system-generated value will be assigned.