Managedblockchain

This page documents function available when using the Managedblockchain module, created with @service Managedblockchain.

Index

Main.Managedblockchain.create_accessor

Main.Managedblockchain.create_member

Main.Managedblockchain.create_network

Main.Managedblockchain.create_node

Main.Managedblockchain.create_proposal

Main.Managedblockchain.delete_accessor

Main.Managedblockchain.delete_member

Main.Managedblockchain.delete_node

Main.Managedblockchain.get_accessor

Main.Managedblockchain.get_member

Main.Managedblockchain.get_network

Main.Managedblockchain.get_node

Main.Managedblockchain.get_proposal

Main.Managedblockchain.list_accessors

Main.Managedblockchain.list_invitations

Main.Managedblockchain.list_members

Main.Managedblockchain.list_networks

Main.Managedblockchain.list_nodes

Main.Managedblockchain.list_proposal_votes

Main.Managedblockchain.list_proposals

Main.Managedblockchain.list_tags_for_resource

Main.Managedblockchain.reject_invitation

Main.Managedblockchain.tag_resource

Main.Managedblockchain.untag_resource

Main.Managedblockchain.update_member

Main.Managedblockchain.update_node

Main.Managedblockchain.vote_on_proposal

Documentation

Main.Managedblockchain.create_accessor — Method

create_accessor(accessor_type, client_request_token)
create_accessor(accessor_type, client_request_token, params::Dict{String,<:Any})

The token based access feature is in preview release for Ethereum on Amazon Managed Blockchain and is subject to change. We recommend that you use this feature only with test scenarios, and not in production environments. Creates a new accessor for use with Managed Blockchain Ethereum nodes. An accessor object is a container that has the information required for token based access to your Ethereum nodes.

Arguments

accessor_type: The type of accessor. Currently accessor type is restricted to BILLING_TOKEN.

client_request_token: This is a unique, case-sensitive identifier that you provide to ensure the idempotency of the operation. An idempotent operation completes no more than once. This identifier is required only if you make a service request directly using an HTTP client. It is generated automatically if you use an Amazon Web Services SDK or the Amazon Web Services CLI.

Main.Managedblockchain.create_member — Method

create_member(client_request_token, invitation_id, member_configuration, network_id)
create_member(client_request_token, invitation_id, member_configuration, network_id, params::Dict{String,<:Any})

Creates a member within a Managed Blockchain network. Applies only to Hyperledger Fabric.

Arguments

client_request_token: A unique, case-sensitive identifier that you provide to ensure the idempotency of the operation. An idempotent operation completes no more than one time. This identifier is required only if you make a service request directly using an HTTP client. It is generated automatically if you use an Amazon Web Services SDK or the CLI.

invitation_id: The unique identifier of the invitation that is sent to the member to join the network.

member_configuration: Member configuration parameters.

network_id: The unique identifier of the network in which the member is created.

Main.Managedblockchain.create_network — Method

create_network(client_request_token, framework, framework_version, member_configuration, name, voting_policy)
create_network(client_request_token, framework, framework_version, member_configuration, name, voting_policy, params::Dict{String,<:Any})

Creates a new blockchain network using Amazon Managed Blockchain. Applies only to Hyperledger Fabric.

Arguments

client_request_token: This is a unique, case-sensitive identifier that you provide to ensure the idempotency of the operation. An idempotent operation completes no more than once. This identifier is required only if you make a service request directly using an HTTP client. It is generated automatically if you use an Amazon Web Services SDK or the Amazon Web Services CLI.

framework: The blockchain framework that the network uses.

framework_version: The version of the blockchain framework that the network uses.

member_configuration: Configuration properties for the first member within the network.

name: The name of the network.

voting_policy: The voting rules used by the network to determine if a proposal is approved.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": An optional description for the network.

"FrameworkConfiguration": Configuration properties of the blockchain framework relevant to the network configuration.

"Tags": Tags to assign to the network. Each tag consists of a key and optional value. When specifying tags during creation, you can specify multiple key-value pairs in a single request, with an overall maximum of 50 tags added to each resource. For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Ethereum Developer Guide, or Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

Main.Managedblockchain.create_node — Method

create_node(client_request_token, node_configuration, network_id)
create_node(client_request_token, node_configuration, network_id, params::Dict{String,<:Any})

Creates a node on the specified blockchain network. Applies to Hyperledger Fabric and Ethereum.

Arguments

client_request_token: A unique, case-sensitive identifier that you provide to ensure the idempotency of the operation. An idempotent operation completes no more than one time. This identifier is required only if you make a service request directly using an HTTP client. It is generated automatically if you use an Amazon Web Services SDK or the CLI.

node_configuration: The properties of a node configuration.

network_id: The unique identifier of the network for the node. Ethereum public networks have the following NetworkIds: n-ethereum-mainnet n-ethereum-goerli n-ethereum-rinkeby n-ethereum-ropsten

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MemberId": The unique identifier of the member that owns this node. Applies only to Hyperledger Fabric.

"Tags": Tags to assign to the node. Each tag consists of a key and optional value. When specifying tags during creation, you can specify multiple key-value pairs in a single request, with an overall maximum of 50 tags added to each resource. For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Ethereum Developer Guide, or Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

Main.Managedblockchain.create_proposal — Method

create_proposal(actions, client_request_token, member_id, network_id)
create_proposal(actions, client_request_token, member_id, network_id, params::Dict{String,<:Any})

Creates a proposal for a change to the network that other members of the network can vote on, for example, a proposal to add a new member to the network. Any member can create a proposal. Applies only to Hyperledger Fabric.

Arguments

actions: The type of actions proposed, such as inviting a member or removing a member. The types of Actions in a proposal are mutually exclusive. For example, a proposal with Invitations actions cannot also contain Removals actions.

client_request_token: A unique, case-sensitive identifier that you provide to ensure the idempotency of the operation. An idempotent operation completes no more than one time. This identifier is required only if you make a service request directly using an HTTP client. It is generated automatically if you use an Amazon Web Services SDK or the CLI.

member_id: The unique identifier of the member that is creating the proposal. This identifier is especially useful for identifying the member making the proposal when multiple members exist in a single Amazon Web Services account.

network_id: The unique identifier of the network for which the proposal is made.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A description for the proposal that is visible to voting members, for example, "Proposal to add Example Corp. as member."

"Tags": Tags to assign to the proposal. Each tag consists of a key and optional value. When specifying tags during creation, you can specify multiple key-value pairs in a single request, with an overall maximum of 50 tags added to each resource. If the proposal is for a network invitation, the invitation inherits the tags added to the proposal. For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Ethereum Developer Guide, or Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

Main.Managedblockchain.delete_accessor — Method

delete_accessor(accessor_id)
delete_accessor(accessor_id, params::Dict{String,<:Any})

The token based access feature is in preview release for Ethereum on Amazon Managed Blockchain and is subject to change. We recommend that you use this feature only with test scenarios, and not in production environments. Deletes an accessor that your Amazon Web Services account owns. An accessor object is a container that has the information required for token based access to your Ethereum nodes including, the BILLINGTOKEN. After an accessor is deleted, the status of the accessor changes from AVAILABLE to PENDINGDELETION. An accessor in the PENDING_DELETION state can’t be used for new WebSocket requests or HTTP requests. However, WebSocket connections that were initiated while the accessor was in the AVAILABLE state remain open until they expire (up to 2 hours).

Arguments

accessor_id: The unique identifier of the accessor.

Main.Managedblockchain.delete_member — Method

delete_member(member_id, network_id)
delete_member(member_id, network_id, params::Dict{String,<:Any})

Deletes a member. Deleting a member removes the member and all associated resources from the network. DeleteMember can only be called for a specified MemberId if the principal performing the action is associated with the Amazon Web Services account that owns the member. In all other cases, the DeleteMember action is carried out as the result of an approved proposal to remove a member. If MemberId is the last member in a network specified by the last Amazon Web Services account, the network is deleted also. Applies only to Hyperledger Fabric.

Arguments

member_id: The unique identifier of the member to remove.

network_id: The unique identifier of the network from which the member is removed.

Main.Managedblockchain.delete_node — Method

delete_node(network_id, node_id)
delete_node(network_id, node_id, params::Dict{String,<:Any})

Deletes a node that your Amazon Web Services account owns. All data on the node is lost and cannot be recovered. Applies to Hyperledger Fabric and Ethereum.

Arguments

network_id: The unique identifier of the network that the node is on. Ethereum public networks have the following NetworkIds: n-ethereum-mainnet n-ethereum-goerli n-ethereum-rinkeby n-ethereum-ropsten

node_id: The unique identifier of the node.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"memberId": The unique identifier of the member that owns this node. Applies only to Hyperledger Fabric and is required for Hyperledger Fabric.

Main.Managedblockchain.get_accessor — Method

get_accessor(accessor_id)
get_accessor(accessor_id, params::Dict{String,<:Any})

The token based access feature is in preview release for Ethereum on Amazon Managed Blockchain and is subject to change. We recommend that you use this feature only with test scenarios, and not in production environments. Returns detailed information about an accessor. An accessor object is a container that has the information required for token based access to your Ethereum nodes.

Arguments

accessor_id: The unique identifier of the accessor.

Main.Managedblockchain.get_member — Method

get_member(member_id, network_id)
get_member(member_id, network_id, params::Dict{String,<:Any})

Returns detailed information about a member. Applies only to Hyperledger Fabric.

Arguments

member_id: The unique identifier of the member.

network_id: The unique identifier of the network to which the member belongs.

Main.Managedblockchain.get_network — Method

get_network(network_id)
get_network(network_id, params::Dict{String,<:Any})

Returns detailed information about a network. Applies to Hyperledger Fabric and Ethereum.

Arguments

network_id: The unique identifier of the network to get information about.

Main.Managedblockchain.get_node — Method

get_node(network_id, node_id)
get_node(network_id, node_id, params::Dict{String,<:Any})

Returns detailed information about a node. Applies to Hyperledger Fabric and Ethereum.

Arguments

network_id: The unique identifier of the network that the node is on.

node_id: The unique identifier of the node.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"memberId": The unique identifier of the member that owns the node. Applies only to Hyperledger Fabric and is required for Hyperledger Fabric.

Main.Managedblockchain.get_proposal — Method

get_proposal(network_id, proposal_id)
get_proposal(network_id, proposal_id, params::Dict{String,<:Any})

Returns detailed information about a proposal. Applies only to Hyperledger Fabric.

Arguments

network_id: The unique identifier of the network for which the proposal is made.

proposal_id: The unique identifier of the proposal.

Main.Managedblockchain.list_accessors — Method

list_accessors()
list_accessors(params::Dict{String,<:Any})

The token based access feature is in preview release for Ethereum on Amazon Managed Blockchain and is subject to change. We recommend that you use this feature only with test scenarios, and not in production environments. Returns a list of the accessors and their properties. Accessor objects are containers that have the information required for token based access to your Ethereum nodes.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of accessors to list.

"nextToken": The pagination token that indicates the next set of results to retrieve.

Main.Managedblockchain.list_invitations — Method

list_invitations()
list_invitations(params::Dict{String,<:Any})

Returns a list of all invitations for the current Amazon Web Services account. Applies only to Hyperledger Fabric.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of invitations to return.

"nextToken": The pagination token that indicates the next set of results to retrieve.

Main.Managedblockchain.list_members — Method

list_members(network_id)
list_members(network_id, params::Dict{String,<:Any})

Returns a list of the members in a network and properties of their configurations. Applies only to Hyperledger Fabric.

Arguments

network_id: The unique identifier of the network for which to list members.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"isOwned": An optional Boolean value. If provided, the request is limited either to members that the current Amazon Web Services account owns (true) or that other Amazon Web Services accountsn own (false). If omitted, all members are listed.

"maxResults": The maximum number of members to return in the request.

"name": The optional name of the member to list.

"nextToken": The pagination token that indicates the next set of results to retrieve.

"status": An optional status specifier. If provided, only members currently in this status are listed.

Main.Managedblockchain.list_networks — Method

list_networks()
list_networks(params::Dict{String,<:Any})

Returns information about the networks in which the current Amazon Web Services account participates. Applies to Hyperledger Fabric and Ethereum.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"framework": An optional framework specifier. If provided, only networks of this framework type are listed.

"maxResults": The maximum number of networks to list.

"name": The name of the network.

"nextToken": The pagination token that indicates the next set of results to retrieve.

"status": An optional status specifier. If provided, only networks currently in this status are listed. Applies only to Hyperledger Fabric.

Main.Managedblockchain.list_nodes — Method

list_nodes(network_id)
list_nodes(network_id, params::Dict{String,<:Any})

Returns information about the nodes within a network. Applies to Hyperledger Fabric and Ethereum.

Arguments

network_id: The unique identifier of the network for which to list nodes.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of nodes to list.

"memberId": The unique identifier of the member who owns the nodes to list. Applies only to Hyperledger Fabric and is required for Hyperledger Fabric.

"nextToken": The pagination token that indicates the next set of results to retrieve.

"status": An optional status specifier. If provided, only nodes currently in this status are listed.

Main.Managedblockchain.list_proposal_votes — Method

list_proposal_votes(network_id, proposal_id)
list_proposal_votes(network_id, proposal_id, params::Dict{String,<:Any})

Returns the list of votes for a specified proposal, including the value of each vote and the unique identifier of the member that cast the vote. Applies only to Hyperledger Fabric.

Arguments

network_id: The unique identifier of the network.

proposal_id: The unique identifier of the proposal.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of votes to return.

"nextToken": The pagination token that indicates the next set of results to retrieve.

Main.Managedblockchain.list_proposals — Method

list_proposals(network_id)
list_proposals(network_id, params::Dict{String,<:Any})

Returns a list of proposals for the network. Applies only to Hyperledger Fabric.

Arguments

network_id: The unique identifier of the network.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of proposals to return.

"nextToken": The pagination token that indicates the next set of results to retrieve.

Main.Managedblockchain.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Returns a list of tags for the specified resource. Each tag consists of a key and optional value. For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Ethereum Developer Guide, or Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

Main.Managedblockchain.reject_invitation — Method

reject_invitation(invitation_id)
reject_invitation(invitation_id, params::Dict{String,<:Any})

Rejects an invitation to join a network. This action can be called by a principal in an Amazon Web Services account that has received an invitation to create a member and join a network. Applies only to Hyperledger Fabric.

Arguments

invitation_id: The unique identifier of the invitation to reject.

Main.Managedblockchain.tag_resource — Method

tag_resource(tags, resource_arn)
tag_resource(tags, resource_arn, params::Dict{String,<:Any})

Adds or overwrites the specified tags for the specified Amazon Managed Blockchain resource. Each tag consists of a key and optional value. When you specify a tag key that already exists, the tag value is overwritten with the new value. Use UntagResource to remove tag keys. A resource can have up to 50 tags. If you try to create more than 50 tags for a resource, your request fails and returns an error. For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Ethereum Developer Guide, or Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

Arguments

tags: The tags to assign to the specified resource. Tag values can be empty, for example, "MyTagKey" : "". You can specify multiple key-value pairs in a single request, with an overall maximum of 50 tags added to each resource.

resource_arn: The Amazon Resource Name (ARN) of the resource. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

Main.Managedblockchain.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes the specified tags from the Amazon Managed Blockchain resource. For more information about tags, see Tagging Resources in the Amazon Managed Blockchain Ethereum Developer Guide, or Tagging Resources in the Amazon Managed Blockchain Hyperledger Fabric Developer Guide.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource. For more information about ARNs and their format, see Amazon Resource Names (ARNs) in the Amazon Web Services General Reference.

tag_keys: The tag keys.

Main.Managedblockchain.update_member — Method

update_member(member_id, network_id)
update_member(member_id, network_id, params::Dict{String,<:Any})

Updates a member configuration with new parameters. Applies only to Hyperledger Fabric.

Arguments

member_id: The unique identifier of the member.

network_id: The unique identifier of the Managed Blockchain network to which the member belongs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"LogPublishingConfiguration": Configuration properties for publishing to Amazon CloudWatch Logs.

Main.Managedblockchain.update_node — Method

update_node(network_id, node_id)
update_node(network_id, node_id, params::Dict{String,<:Any})

Updates a node configuration with new parameters. Applies only to Hyperledger Fabric.

Arguments

network_id: The unique identifier of the network that the node is on.

node_id: The unique identifier of the node.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"LogPublishingConfiguration": Configuration properties for publishing to Amazon CloudWatch Logs.

"MemberId": The unique identifier of the member that owns the node. Applies only to Hyperledger Fabric.

Main.Managedblockchain.vote_on_proposal — Method

vote_on_proposal(vote, voter_member_id, network_id, proposal_id)
vote_on_proposal(vote, voter_member_id, network_id, proposal_id, params::Dict{String,<:Any})

Casts a vote for a specified ProposalId on behalf of a member. The member to vote as, specified by VoterMemberId, must be in the same Amazon Web Services account as the principal that calls the action. Applies only to Hyperledger Fabric.

Arguments

vote: The value of the vote.

voter_member_id: The unique identifier of the member casting the vote.

network_id: The unique identifier of the network.

proposal_id: The unique identifier of the proposal.