Marketplace Catalog

This page documents function available when using the Marketplace_Catalog module, created with @service Marketplace_Catalog.

Index

Main.Marketplace_Catalog.cancel_change_set

Main.Marketplace_Catalog.describe_change_set

Main.Marketplace_Catalog.describe_entity

Main.Marketplace_Catalog.list_change_sets

Main.Marketplace_Catalog.list_entities

Main.Marketplace_Catalog.start_change_set

Documentation

Main.Marketplace_Catalog.cancel_change_set — Method

cancel_change_set(catalog, change_set_id)
cancel_change_set(catalog, change_set_id, params::Dict{String,<:Any})

Used to cancel an open change request. Must be sent before the status of the request changes to APPLYING, the final stage of completing your change request. You can describe a change during the 60-day request history retention period for API calls.

Arguments

catalog: Required. The catalog related to the request. Fixed value: AWSMarketplace.

change_set_id: Required. The unique identifier of the StartChangeSet request that you want to cancel.

Main.Marketplace_Catalog.describe_change_set — Method

describe_change_set(catalog, change_set_id)
describe_change_set(catalog, change_set_id, params::Dict{String,<:Any})

Provides information about a given change set.

Arguments

catalog: Required. The catalog related to the request. Fixed value: AWSMarketplace

change_set_id: Required. The unique identifier for the StartChangeSet request that you want to describe the details for.

Main.Marketplace_Catalog.describe_entity — Method

describe_entity(catalog, entity_id)
describe_entity(catalog, entity_id, params::Dict{String,<:Any})

Returns the metadata and content of the entity.

Arguments

catalog: Required. The catalog related to the request. Fixed value: AWSMarketplace

entity_id: Required. The unique ID of the entity to describe.

Main.Marketplace_Catalog.list_change_sets — Method

list_change_sets(catalog)
list_change_sets(catalog, params::Dict{String,<:Any})

Returns the list of change sets owned by the account being used to make the call. You can filter this list by providing any combination of entityId, ChangeSetName, and status. If you provide more than one filter, the API operation applies a logical AND between the filters. You can describe a change during the 60-day request history retention period for API calls.

Arguments

catalog: The catalog related to the request. Fixed value: AWSMarketplace

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"FilterList": An array of filter objects.

"MaxResults": The maximum number of results returned by a single call. This value must be provided in the next call to retrieve the next set of results. By default, this value is

"NextToken": The token value retrieved from a previous call to access the next page of results.

"Sort": An object that contains two attributes, SortBy and SortOrder.

Main.Marketplace_Catalog.list_entities — Method

list_entities(catalog, entity_type)
list_entities(catalog, entity_type, params::Dict{String,<:Any})

Provides the list of entities of a given type.

Arguments

catalog: The catalog related to the request. Fixed value: AWSMarketplace

entity_type: The type of entities to retrieve.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"FilterList": An array of filter objects. Each filter object contains two attributes, filterName and filterValues.

"MaxResults": Specifies the upper limit of the elements on a single page. If a value isn't provided, the default value is 20.

"NextToken": The value of the next token, if it exists. Null if there are no more results.

"Sort": An object that contains two attributes, SortBy and SortOrder.

Main.Marketplace_Catalog.start_change_set — Method

start_change_set(catalog, change_set)
start_change_set(catalog, change_set, params::Dict{String,<:Any})

This operation allows you to request changes for your entities. Within a single ChangeSet, you cannot start the same change type against the same entity multiple times. Additionally, when a ChangeSet is running, all the entities targeted by the different changes are locked until the ChangeSet has completed (either succeeded, cancelled, or failed). If you try to start a ChangeSet containing a change against an entity that is already locked, you will receive a ResourceInUseException. For example, you cannot start the ChangeSet described in the example later in this topic, because it contains two changes to execute the same change type (AddRevisions) against the same entity (entity-id@1). For more information about working with change sets, see Working with change sets.

Arguments

catalog: The catalog related to the request. Fixed value: AWSMarketplace

change_set: Array of change object.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ChangeSetName": Optional case sensitive string of up to 100 ASCII characters. The change set name can be used to filter the list of change sets.

"ClientRequestToken": A unique token to identify the request to ensure idempotency.