Medialive

This page documents function available when using the Medialive module, created with @service Medialive.

Index

Main.Medialive.accept_input_device_transfer

Main.Medialive.batch_delete

Main.Medialive.batch_start

Main.Medialive.batch_stop

Main.Medialive.batch_update_schedule

Main.Medialive.cancel_input_device_transfer

Main.Medialive.create_channel

Main.Medialive.create_input

Main.Medialive.create_input_security_group

Main.Medialive.create_multiplex

Main.Medialive.create_multiplex_program

Main.Medialive.create_tags

Main.Medialive.delete_channel

Main.Medialive.delete_input

Main.Medialive.delete_input_security_group

Main.Medialive.delete_multiplex

Main.Medialive.delete_multiplex_program

Main.Medialive.delete_reservation

Main.Medialive.delete_schedule

Main.Medialive.delete_tags

Main.Medialive.describe_channel

Main.Medialive.describe_input

Main.Medialive.describe_input_device

Main.Medialive.describe_input_device_thumbnail

Main.Medialive.describe_input_security_group

Main.Medialive.describe_multiplex

Main.Medialive.describe_multiplex_program

Main.Medialive.describe_offering

Main.Medialive.describe_reservation

Main.Medialive.describe_schedule

Main.Medialive.list_channels

Main.Medialive.list_input_device_transfers

Main.Medialive.list_input_devices

Main.Medialive.list_input_security_groups

Main.Medialive.list_inputs

Main.Medialive.list_multiplex_programs

Main.Medialive.list_multiplexes

Main.Medialive.list_offerings

Main.Medialive.list_reservations

Main.Medialive.list_tags_for_resource

Main.Medialive.purchase_offering

Main.Medialive.reject_input_device_transfer

Main.Medialive.start_channel

Main.Medialive.start_multiplex

Main.Medialive.stop_channel

Main.Medialive.stop_multiplex

Main.Medialive.transfer_input_device

Main.Medialive.update_channel

Main.Medialive.update_channel_class

Main.Medialive.update_input

Main.Medialive.update_input_device

Main.Medialive.update_input_security_group

Main.Medialive.update_multiplex

Main.Medialive.update_multiplex_program

Main.Medialive.update_reservation

Documentation

Main.Medialive.accept_input_device_transfer — Method

AcceptInputDeviceTransfer()

Accept an incoming input device transfer. The ownership of the device will transfer to your AWS account.

Required Parameters

inputDeviceId: The unique ID of the input device to accept. For example, hd-123456789abcdef.

Main.Medialive.batch_delete — Method

BatchDelete()

Starts delete of resources.

Optional Parameters

channelIds: List of channel IDs

inputIds: List of input IDs

inputSecurityGroupIds: List of input security group IDs

multiplexIds: List of multiplex IDs

Main.Medialive.batch_start — Method

BatchStart()

Starts existing resources

Optional Parameters

channelIds: List of channel IDs

multiplexIds: List of multiplex IDs

Main.Medialive.batch_stop — Method

BatchStop()

Stops running resources

Optional Parameters

channelIds: List of channel IDs

multiplexIds: List of multiplex IDs

Main.Medialive.batch_update_schedule — Method

BatchUpdateSchedule()

Update a channel schedule

Required Parameters

channelId: Id of the channel whose schedule is being updated.

Optional Parameters

creates: Schedule actions to create in the schedule.

deletes: Schedule actions to delete from the schedule.

Main.Medialive.cancel_input_device_transfer — Method

CancelInputDeviceTransfer()

Cancel an input device transfer that you have requested.

Required Parameters

inputDeviceId: The unique ID of the input device to cancel. For example, hd-123456789abcdef.

Main.Medialive.create_channel — Method

CreateChannel()

Creates a new channel

Optional Parameters

cdiInputSpecification: Specification of CDI inputs for this channel

channelClass: The class for this channel. STANDARD for a channel with two pipelines or SINGLE_PIPELINE for a channel with one pipeline.

destinations:

encoderSettings:

inputAttachments: List of input attachments for channel.

inputSpecification: Specification of network and file inputs for this channel

logLevel: The log level to write to CloudWatch Logs.

name: Name of channel.

requestId: Unique request ID to be specified. This is needed to prevent retries from

creating multiple resources.

reserved: Deprecated field that's only usable by whitelisted customers.

roleArn: An optional Amazon Resource Name (ARN) of the role to assume when running the Channel.

tags: A collection of key-value pairs.

Main.Medialive.create_input — Method

CreateInput()

Create an input

Optional Parameters

destinations: Destination settings for PUSH type inputs.

inputDevices: Settings for the devices.

inputSecurityGroups: A list of security groups referenced by IDs to attach to the input.

mediaConnectFlows: A list of the MediaConnect Flows that you want to use in this input. You can specify as few as one

Flow and presently, as many as two. The only requirement is when you have more than one is that each Flow is in a separate Availability Zone as this ensures your EML input is redundant to AZ issues.

name: Name of the input.

requestId: Unique identifier of the request to ensure the request is handled

exactly once in case of retries.

roleArn: The Amazon Resource Name (ARN) of the role this input assumes during and after creation.

sources: The source URLs for a PULL-type input. Every PULL type input needs

exactly two source URLs for redundancy. Only specify sources for PULL type Inputs. Leave Destinations empty.

tags: A collection of key-value pairs.

type:

vpc:

Main.Medialive.create_input_security_group — Method

CreateInputSecurityGroup()

Creates a Input Security Group

Optional Parameters

tags: A collection of key-value pairs.

whitelistRules: List of IPv4 CIDR addresses to whitelist

Main.Medialive.create_multiplex — Method

CreateMultiplex()

Create a new multiplex.

Required Parameters

availabilityZones: A list of availability zones for the multiplex. You must specify exactly two.

multiplexSettings: Configuration for a multiplex event.

name: Name of multiplex.

requestId: Unique request ID. This prevents retries from creating multiple

resources.

Optional Parameters

tags: A collection of key-value pairs.

Main.Medialive.create_multiplex_program — Method

CreateMultiplexProgram()

Create a new program in the multiplex.

Required Parameters

multiplexId: ID of the multiplex where the program is to be created.

multiplexProgramSettings: The settings for this multiplex program.

programName: Name of multiplex program.

requestId: Unique request ID. This prevents retries from creating multiple

resources.

Main.Medialive.create_tags — Method

CreateTags()

Create tags for a resource

Required Parameters

resource-arn:

Optional Parameters

tags:

Main.Medialive.delete_channel — Method

DeleteChannel()

Starts deletion of channel. The associated outputs are also deleted.

Required Parameters

channelId: Unique ID of the channel.

Main.Medialive.delete_input — Method

DeleteInput()

Deletes the input end point

Required Parameters

inputId: Unique ID of the input

Main.Medialive.delete_input_security_group — Method

DeleteInputSecurityGroup()

Deletes an Input Security Group

Required Parameters

inputSecurityGroupId: The Input Security Group to delete

Main.Medialive.delete_multiplex — Method

DeleteMultiplex()

Delete a multiplex. The multiplex must be idle.

Required Parameters

multiplexId: The ID of the multiplex.

Main.Medialive.delete_multiplex_program — Method

DeleteMultiplexProgram()

Delete a program from a multiplex.

Required Parameters

multiplexId: The ID of the multiplex that the program belongs to.

programName: The multiplex program name.

Main.Medialive.delete_reservation — Method

DeleteReservation()

Delete an expired reservation.

Required Parameters

reservationId: Unique reservation ID, e.g. '1234567'

Main.Medialive.delete_schedule — Method

DeleteSchedule()

Delete all schedule actions on a channel.

Required Parameters

channelId: Id of the channel whose schedule is being deleted.

Main.Medialive.delete_tags — Method

DeleteTags()

Removes tags for a resource

Required Parameters

resource-arn:

tagKeys: An array of tag keys to delete

Main.Medialive.describe_channel — Method

DescribeChannel()

Gets details about a channel

Required Parameters

channelId: channel ID

Main.Medialive.describe_input — Method

DescribeInput()

Produces details about an input

Required Parameters

inputId: Unique ID of the input

Main.Medialive.describe_input_device — Method

DescribeInputDevice()

Gets the details for the input device

Required Parameters

inputDeviceId: The unique ID of this input device. For example, hd-123456789abcdef.

Main.Medialive.describe_input_device_thumbnail — Method

DescribeInputDeviceThumbnail()

Get the latest thumbnail data for the input device.

Required Parameters

accept: The HTTP Accept header. Indicates the requested type for the thumbnail.

inputDeviceId: The unique ID of this input device. For example, hd-123456789abcdef.

Main.Medialive.describe_input_security_group — Method

DescribeInputSecurityGroup()

Produces a summary of an Input Security Group

Required Parameters

inputSecurityGroupId: The id of the Input Security Group to describe

Main.Medialive.describe_multiplex — Method

DescribeMultiplex()

Gets details about a multiplex.

Required Parameters

multiplexId: The ID of the multiplex.

Main.Medialive.describe_multiplex_program — Method

DescribeMultiplexProgram()

Get the details for a program in a multiplex.

Required Parameters

multiplexId: The ID of the multiplex that the program belongs to.

programName: The name of the program.

Main.Medialive.describe_offering — Method

DescribeOffering()

Get details for an offering.

Required Parameters

offeringId: Unique offering ID, e.g. '87654321'

Main.Medialive.describe_reservation — Method

DescribeReservation()

Get details for a reservation.

Required Parameters

reservationId: Unique reservation ID, e.g. '1234567'

Main.Medialive.describe_schedule — Method

DescribeSchedule()

Get a channel schedule

Required Parameters

channelId: Id of the channel whose schedule is being updated.

Optional Parameters

maxResults:

nextToken:

Main.Medialive.list_channels — Method

ListChannels()

Produces list of channels that have been created

Optional Parameters

maxResults:

nextToken:

Main.Medialive.list_input_device_transfers — Method

ListInputDeviceTransfers()

List input devices that are currently being transferred. List input devices that you are transferring from your AWS account or input devices that another AWS account is transferring to you.

Required Parameters

transferType:

Optional Parameters

maxResults:

nextToken:

Main.Medialive.list_input_devices — Method

ListInputDevices()

List input devices

Optional Parameters

maxResults:

nextToken:

Main.Medialive.list_input_security_groups — Method

ListInputSecurityGroups()

Produces a list of Input Security Groups for an account

Optional Parameters

maxResults:

nextToken:

Main.Medialive.list_inputs — Method

ListInputs()

Produces list of inputs that have been created

Optional Parameters

maxResults:

nextToken:

Main.Medialive.list_multiplex_programs — Method

ListMultiplexPrograms()

List the programs that currently exist for a specific multiplex.

Required Parameters

multiplexId: The ID of the multiplex that the programs belong to.

Optional Parameters

maxResults: The maximum number of items to return.

nextToken: The token to retrieve the next page of results.

Main.Medialive.list_multiplexes — Method

ListMultiplexes()

Retrieve a list of the existing multiplexes.

Optional Parameters

maxResults: The maximum number of items to return.

nextToken: The token to retrieve the next page of results.

Main.Medialive.list_offerings — Method

ListOfferings()

List offerings available for purchase.

Optional Parameters

channelClass: Filter by channel class, 'STANDARD' or 'SINGLE_PIPELINE'

channelConfiguration: Filter to offerings that match the configuration of an existing channel, e.g. '2345678' (a channel ID)

codec: Filter by codec, 'AVC', 'HEVC', 'MPEG2', 'AUDIO', or 'LINK'

duration: Filter by offering duration, e.g. '12'

maxResults:

maximumBitrate: Filter by bitrate, 'MAX10MBPS', 'MAX20MBPS', or 'MAX50MBPS'

maximumFramerate: Filter by framerate, 'MAX30FPS' or 'MAX60FPS'

nextToken:

resolution: Filter by resolution, 'SD', 'HD', 'FHD', or 'UHD'

resourceType: Filter by resource type, 'INPUT', 'OUTPUT', 'MULTIPLEX', or 'CHANNEL'

specialFeature: Filter by special feature, 'ADVANCEDAUDIO' or 'AUDIONORMALIZATION'

videoQuality: Filter by video quality, 'STANDARD', 'ENHANCED', or 'PREMIUM'

Main.Medialive.list_reservations — Method

ListReservations()

List purchased reservations.

Optional Parameters

channelClass: Filter by channel class, 'STANDARD' or 'SINGLE_PIPELINE'

codec: Filter by codec, 'AVC', 'HEVC', 'MPEG2', 'AUDIO', or 'LINK'

maxResults:

maximumBitrate: Filter by bitrate, 'MAX10MBPS', 'MAX20MBPS', or 'MAX50MBPS'

maximumFramerate: Filter by framerate, 'MAX30FPS' or 'MAX60FPS'

nextToken:

resolution: Filter by resolution, 'SD', 'HD', 'FHD', or 'UHD'

resourceType: Filter by resource type, 'INPUT', 'OUTPUT', 'MULTIPLEX', or 'CHANNEL'

specialFeature: Filter by special feature, 'ADVANCEDAUDIO' or 'AUDIONORMALIZATION'

videoQuality: Filter by video quality, 'STANDARD', 'ENHANCED', or 'PREMIUM'

Main.Medialive.list_tags_for_resource — Method

ListTagsForResource()

Produces list of tags that have been created for a resource

Required Parameters

resource-arn:

Main.Medialive.purchase_offering — Method

PurchaseOffering()

Purchase an offering and create a reservation.

Required Parameters

count: Number of resources

offeringId: Offering to purchase, e.g. '87654321'

Optional Parameters

name: Name for the new reservation

requestId: Unique request ID to be specified. This is needed to prevent retries from creating multiple resources.

start: Requested reservation start time (UTC) in ISO-8601 format. The specified time must be between the first day of the current month and one year from now. If no value is given, the default is now.

tags: A collection of key-value pairs

Main.Medialive.reject_input_device_transfer — Method

RejectInputDeviceTransfer()

Reject the transfer of the specified input device to your AWS account.

Required Parameters

inputDeviceId: The unique ID of the input device to reject. For example, hd-123456789abcdef.

Main.Medialive.start_channel — Method

StartChannel()

Starts an existing channel

Required Parameters

channelId: A request to start a channel

Main.Medialive.start_multiplex — Method

StartMultiplex()

Start (run) the multiplex. Starting the multiplex does not start the channels. You must explicitly start each channel.

Required Parameters

multiplexId: The ID of the multiplex.

Main.Medialive.stop_channel — Method

StopChannel()

Stops a running channel

Required Parameters

channelId: A request to stop a running channel

Main.Medialive.stop_multiplex — Method

StopMultiplex()

Stops a running multiplex. If the multiplex isn't running, this action has no effect.

Required Parameters

multiplexId: The ID of the multiplex.

Main.Medialive.transfer_input_device — Method

TransferInputDevice()

Start an input device transfer to another AWS account. After you make the request, the other account must accept or reject the transfer.

Required Parameters

inputDeviceId: The unique ID of this input device. For example, hd-123456789abcdef.

Optional Parameters

targetCustomerId: The AWS account ID (12 digits) for the recipient of the device transfer.

transferMessage: An optional message for the recipient. Maximum 280 characters.

Main.Medialive.update_channel — Method

UpdateChannel()

Updates a channel.

Required Parameters

channelId: channel ID

Optional Parameters

cdiInputSpecification: Specification of CDI inputs for this channel

destinations: A list of output destinations for this channel.

encoderSettings: The encoder settings for this channel.

inputAttachments:

inputSpecification: Specification of network and file inputs for this channel

logLevel: The log level to write to CloudWatch Logs.

name: The name of the channel.

roleArn: An optional Amazon Resource Name (ARN) of the role to assume when running the Channel. If you do not specify this on an update call but the role was previously set that role will be removed.

Main.Medialive.update_channel_class — Method

UpdateChannelClass()

Changes the class of the channel.

Required Parameters

channelClass: The channel class that you wish to update this channel to use.

channelId: Channel Id of the channel whose class should be updated.

Optional Parameters

destinations: A list of output destinations for this channel.

Main.Medialive.update_input — Method

UpdateInput()

Updates an input.

Required Parameters

inputId: Unique ID of the input.

Optional Parameters

destinations: Destination settings for PUSH type inputs.

inputDevices: Settings for the devices.

inputSecurityGroups: A list of security groups referenced by IDs to attach to the input.

mediaConnectFlows: A list of the MediaConnect Flow ARNs that you want to use as the source of the input. You can specify as few as one

Flow and presently, as many as two. The only requirement is when you have more than one is that each Flow is in a separate Availability Zone as this ensures your EML input is redundant to AZ issues.

name: Name of the input.

roleArn: The Amazon Resource Name (ARN) of the role this input assumes during and after creation.

sources: The source URLs for a PULL-type input. Every PULL type input needs

exactly two source URLs for redundancy. Only specify sources for PULL type Inputs. Leave Destinations empty.

Main.Medialive.update_input_device — Method

UpdateInputDevice()

Updates the parameters for the input device.

Required Parameters

inputDeviceId: The unique ID of the input device. For example, hd-123456789abcdef.

Optional Parameters

hdDeviceSettings: The settings that you want to apply to the HD input device.

name: The name that you assigned to this input device (not the unique ID).

uhdDeviceSettings: The settings that you want to apply to the UHD input device.

Main.Medialive.update_input_security_group — Method

UpdateInputSecurityGroup()

Update an Input Security Group's Whilelists.

Required Parameters

inputSecurityGroupId: The id of the Input Security Group to update.

Optional Parameters

tags: A collection of key-value pairs.

whitelistRules: List of IPv4 CIDR addresses to whitelist

Main.Medialive.update_multiplex — Method

UpdateMultiplex()

Updates a multiplex.

Required Parameters

multiplexId: ID of the multiplex to update.

Optional Parameters

multiplexSettings: The new settings for a multiplex.

name: Name of the multiplex.

Main.Medialive.update_multiplex_program — Method

UpdateMultiplexProgram()

Update a program in a multiplex.

Required Parameters

multiplexId: The ID of the multiplex of the program to update.

programName: The name of the program to update.

Optional Parameters

multiplexProgramSettings: The new settings for a multiplex program.

Main.Medialive.update_reservation — Method

UpdateReservation()

Update reservation.

Required Parameters

reservationId: Unique reservation ID, e.g. '1234567'

Optional Parameters

name: Name of the reservation