Migrationhuborchestrator

This page documents function available when using the Migrationhuborchestrator module, created with @service Migrationhuborchestrator.

Index

Main.Migrationhuborchestrator.create_workflow

Main.Migrationhuborchestrator.create_workflow_step

Main.Migrationhuborchestrator.create_workflow_step_group

Main.Migrationhuborchestrator.delete_workflow

Main.Migrationhuborchestrator.delete_workflow_step

Main.Migrationhuborchestrator.delete_workflow_step_group

Main.Migrationhuborchestrator.get_template

Main.Migrationhuborchestrator.get_template_step

Main.Migrationhuborchestrator.get_template_step_group

Main.Migrationhuborchestrator.get_workflow

Main.Migrationhuborchestrator.get_workflow_step

Main.Migrationhuborchestrator.get_workflow_step_group

Main.Migrationhuborchestrator.list_plugins

Main.Migrationhuborchestrator.list_tags_for_resource

Main.Migrationhuborchestrator.list_template_step_groups

Main.Migrationhuborchestrator.list_template_steps

Main.Migrationhuborchestrator.list_templates

Main.Migrationhuborchestrator.list_workflow_step_groups

Main.Migrationhuborchestrator.list_workflow_steps

Main.Migrationhuborchestrator.list_workflows

Main.Migrationhuborchestrator.retry_workflow_step

Main.Migrationhuborchestrator.start_workflow

Main.Migrationhuborchestrator.stop_workflow

Main.Migrationhuborchestrator.tag_resource

Main.Migrationhuborchestrator.untag_resource

Main.Migrationhuborchestrator.update_workflow

Main.Migrationhuborchestrator.update_workflow_step

Main.Migrationhuborchestrator.update_workflow_step_group

Documentation

Main.Migrationhuborchestrator.create_workflow — Method

create_workflow(application_configuration_id, input_parameters, name, template_id)
create_workflow(application_configuration_id, input_parameters, name, template_id, params::Dict{String,<:Any})

Create a workflow to orchestrate your migrations.

Arguments

application_configuration_id: The configuration ID of the application configured in Application Discovery Service.

input_parameters: The input parameters required to create a migration workflow.

name: The name of the migration workflow.

template_id: The ID of the template.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The description of the migration workflow.

"stepTargets": The servers on which a step will be run.

"tags": The tags to add on a migration workflow.

Main.Migrationhuborchestrator.create_workflow_step — Method

create_workflow_step(name, step_action_type, step_group_id, workflow_id)
create_workflow_step(name, step_action_type, step_group_id, workflow_id, params::Dict{String,<:Any})

Create a step in the migration workflow.

Arguments

name: The name of the step.

step_action_type: The action type of the step. You must run and update the status of a manual step for the workflow to continue after the completion of the step.

step_group_id: The ID of the step group.

workflow_id: The ID of the migration workflow.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The description of the step.

"next": The next step.

"outputs": The key value pairs added for the expected output.

"previous": The previous step.

"stepTarget": The servers on which a step will be run.

"workflowStepAutomationConfiguration": The custom script to run tests on source or target environments.

Main.Migrationhuborchestrator.create_workflow_step_group — Method

create_workflow_step_group(name, workflow_id)
create_workflow_step_group(name, workflow_id, params::Dict{String,<:Any})

Create a step group in a migration workflow.

Arguments

name: The name of the step group.

workflow_id: The ID of the migration workflow that will contain the step group.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The description of the step group.

"next": The next step group.

"previous": The previous step group.

Main.Migrationhuborchestrator.delete_workflow — Method

delete_workflow(id)
delete_workflow(id, params::Dict{String,<:Any})

Delete a migration workflow. You must pause a running workflow in Migration Hub Orchestrator console to delete it.

Arguments

id: The ID of the migration workflow you want to delete.

Main.Migrationhuborchestrator.delete_workflow_step — Method

delete_workflow_step(id, step_group_id, workflow_id)
delete_workflow_step(id, step_group_id, workflow_id, params::Dict{String,<:Any})

Delete a step in a migration workflow. Pause the workflow to delete a running step.

Arguments

id: The ID of the step you want to delete.

step_group_id: The ID of the step group that contains the step you want to delete.

workflow_id: The ID of the migration workflow.

Main.Migrationhuborchestrator.delete_workflow_step_group — Method

delete_workflow_step_group(id, workflow_id)
delete_workflow_step_group(id, workflow_id, params::Dict{String,<:Any})

Delete a step group in a migration workflow.

Arguments

id: The ID of the step group you want to delete.

workflow_id: The ID of the migration workflow.

Main.Migrationhuborchestrator.get_template — Method

get_template(id)
get_template(id, params::Dict{String,<:Any})

Get the template you want to use for creating a migration workflow.

Arguments

id: The ID of the template.

Main.Migrationhuborchestrator.get_template_step — Method

get_template_step(id, step_group_id, template_id)
get_template_step(id, step_group_id, template_id, params::Dict{String,<:Any})

Get a specific step in a template.

Arguments

id: The ID of the step.

step_group_id: The ID of the step group.

template_id: The ID of the template.

Main.Migrationhuborchestrator.get_template_step_group — Method

get_template_step_group(id, template_id)
get_template_step_group(id, template_id, params::Dict{String,<:Any})

Get a step group in a template.

Arguments

id: The ID of the step group.

template_id: The ID of the template.

Main.Migrationhuborchestrator.get_workflow — Method

get_workflow(id)
get_workflow(id, params::Dict{String,<:Any})

Get migration workflow.

Arguments

id: The ID of the migration workflow.

Main.Migrationhuborchestrator.get_workflow_step — Method

get_workflow_step(id, step_group_id, workflow_id)
get_workflow_step(id, step_group_id, workflow_id, params::Dict{String,<:Any})

Get a step in the migration workflow.

Arguments

id: The ID of the step.

step_group_id: desThe ID of the step group.

workflow_id: The ID of the migration workflow.

Main.Migrationhuborchestrator.get_workflow_step_group — Method

get_workflow_step_group(id, workflow_id)
get_workflow_step_group(id, workflow_id, params::Dict{String,<:Any})

Get the step group of a migration workflow.

Arguments

id: The ID of the step group.

workflow_id: The ID of the migration workflow.

Main.Migrationhuborchestrator.list_plugins — Method

list_plugins()
list_plugins(params::Dict{String,<:Any})

List AWS Migration Hub Orchestrator plugins.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of plugins that can be returned.

"nextToken": The pagination token.

Main.Migrationhuborchestrator.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

List the tags added to a resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

Main.Migrationhuborchestrator.list_template_step_groups — Method

list_template_step_groups(template_id)
list_template_step_groups(template_id, params::Dict{String,<:Any})

List the step groups in a template.

Arguments

template_id: The ID of the template.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results that can be returned.

"nextToken": The pagination token.

Main.Migrationhuborchestrator.list_template_steps — Method

list_template_steps(step_group_id, template_id)
list_template_steps(step_group_id, template_id, params::Dict{String,<:Any})

List the steps in a template.

Arguments

step_group_id: The ID of the step group.

template_id: The ID of the template.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results that can be returned.

"nextToken": The pagination token.

Main.Migrationhuborchestrator.list_templates — Method

list_templates()
list_templates(params::Dict{String,<:Any})

List the templates available in Migration Hub Orchestrator to create a migration workflow.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results that can be returned.

"name": The name of the template.

"nextToken": The pagination token.

Main.Migrationhuborchestrator.list_workflow_step_groups — Method

list_workflow_step_groups(workflow_id)
list_workflow_step_groups(workflow_id, params::Dict{String,<:Any})

List the step groups in a migration workflow.

Arguments

workflow_id: The ID of the migration workflow.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results that can be returned.

"nextToken": The pagination token.

Main.Migrationhuborchestrator.list_workflow_steps — Method

list_workflow_steps(step_group_id, workflow_id)
list_workflow_steps(step_group_id, workflow_id, params::Dict{String,<:Any})

List the steps in a workflow.

Arguments

step_group_id: The ID of the step group.

workflow_id: The ID of the migration workflow.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results that can be returned.

"nextToken": The pagination token.

Main.Migrationhuborchestrator.list_workflows — Method

list_workflows()
list_workflows(params::Dict{String,<:Any})

List the migration workflows.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"adsApplicationConfigurationName": The name of the application configured in Application Discovery Service.

"maxResults": The maximum number of results that can be returned.

"name": The name of the migration workflow.

"nextToken": The pagination token.

"status": The status of the migration workflow.

"templateId": The ID of the template.

Main.Migrationhuborchestrator.retry_workflow_step — Method

retry_workflow_step(id, step_group_id, workflow_id)
retry_workflow_step(id, step_group_id, workflow_id, params::Dict{String,<:Any})

Retry a failed step in a migration workflow.

Arguments

id: The ID of the step.

step_group_id: The ID of the step group.

workflow_id: The ID of the migration workflow.

Main.Migrationhuborchestrator.start_workflow — Method

start_workflow(id)
start_workflow(id, params::Dict{String,<:Any})

Start a migration workflow.

Arguments

id: The ID of the migration workflow.

Main.Migrationhuborchestrator.stop_workflow — Method

stop_workflow(id)
stop_workflow(id, params::Dict{String,<:Any})

Stop an ongoing migration workflow.

Arguments

id: The ID of the migration workflow.

Main.Migrationhuborchestrator.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Tag a resource by specifying its Amazon Resource Name (ARN).

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource to which you want to add tags.

tags: A collection of labels, in the form of key:value pairs, that apply to this resource.

Main.Migrationhuborchestrator.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Deletes the tags for a resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource from which you want to remove tags.

tag_keys: One or more tag keys. Specify only the tag keys, not the tag values.

Main.Migrationhuborchestrator.update_workflow — Method

update_workflow(id)
update_workflow(id, params::Dict{String,<:Any})

Update a migration workflow.

Arguments

id: The ID of the migration workflow.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The description of the migration workflow.

"inputParameters": The input parameters required to update a migration workflow.

"name": The name of the migration workflow.

"stepTargets": The servers on which a step will be run.

Main.Migrationhuborchestrator.update_workflow_step — Method

update_workflow_step(id, step_group_id, workflow_id)
update_workflow_step(id, step_group_id, workflow_id, params::Dict{String,<:Any})

Update a step in a migration workflow.

Arguments

id: The ID of the step.

step_group_id: The ID of the step group.

workflow_id: The ID of the migration workflow.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The description of the step.

"name": The name of the step.

"next": The next step.

"outputs": The outputs of a step.

"previous": The previous step.

"status": The status of the step.

"stepActionType": The action type of the step. You must run and update the status of a manual step for the workflow to continue after the completion of the step.

"stepTarget": The servers on which a step will be run.

"workflowStepAutomationConfiguration": The custom script to run tests on the source and target environments.

Main.Migrationhuborchestrator.update_workflow_step_group — Method

update_workflow_step_group(id, workflow_id)
update_workflow_step_group(id, workflow_id, params::Dict{String,<:Any})

Update the step group in a migration workflow.

Arguments

id: The ID of the step group.

workflow_id: The ID of the migration workflow.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The description of the step group.

"name": The name of the step group.

"next": The next step group.

"previous": The previous step group.