Mpa

This page documents function available when using the Mpa module, created with @service Mpa.

Index

Main.Mpa.cancel_session

Main.Mpa.create_approval_team

Main.Mpa.create_identity_source

Main.Mpa.delete_identity_source

Main.Mpa.delete_inactive_approval_team_version

Main.Mpa.get_approval_team

Main.Mpa.get_identity_source

Main.Mpa.get_policy_version

Main.Mpa.get_resource_policy

Main.Mpa.get_session

Main.Mpa.list_approval_teams

Main.Mpa.list_identity_sources

Main.Mpa.list_policies

Main.Mpa.list_policy_versions

Main.Mpa.list_resource_policies

Main.Mpa.list_sessions

Main.Mpa.list_tags_for_resource

Main.Mpa.start_active_approval_team_deletion

Main.Mpa.start_approval_team_baseline

Main.Mpa.tag_resource

Main.Mpa.untag_resource

Main.Mpa.update_approval_team

Documentation

Main.Mpa.cancel_session — Function

cancel_session(session_arn)
cancel_session(session_arn, params::Dict{String,<:Any})

Cancels an approval session. For more information, see Session in the Multi-party approval User Guide.

Arguments

session_arn: Amazon Resource Name (ARN) for the session.

Main.Mpa.create_approval_team — Function

create_approval_team(approval_strategy, approvers, description, name, policies)
create_approval_team(approval_strategy, approvers, description, name, policies, params::Dict{String,<:Any})

Creates a new approval team. For more information, see Approval team in the Multi-party approval User Guide.

Arguments

approval_strategy: An ApprovalStrategy object. Contains details for how the team grants approval.

approvers: An array of ApprovalTeamRequesterApprovers objects. Contains details for the approvers in the team.

description: Description for the team.

name: Name of the team.

policies: An array of PolicyReference objects. Contains a list of policies that define the permissions for team resources.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If not provided, the Amazon Web Services populates this field.

Note

What is idempotency?

When you make a mutating API request, the request typically returns a result before the operation's asynchronous workflows have completed. Operations might also time out or encounter other server issues before they complete, even though the request has already returned a result. This could make it difficult to determine whether the request succeeded or not, and could lead to multiple retries to ensure that the operation completes successfully. However, if the original request and the subsequent retries are successful, the operation is completed multiple times. This means that you might create more resources than you intended.

Idempotency ensures that an API request completes no more than one time. With an idempotent request, if the original request completes successfully, any subsequent retries complete successfully without performing any further actions.

"Tags": Tags you want to attach to the team.

Main.Mpa.create_identity_source — Function

create_identity_source(identity_source_parameters)
create_identity_source(identity_source_parameters, params::Dict{String,<:Any})

Creates a new identity source. For more information, see Identity Source in the Multi-party approval User Guide.

Arguments

identity_source_parameters: A IdentitySourceParameters object. Contains details for the resource that provides identities to the identity source. For example, an IAM Identity Center instance.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If not provided, the Amazon Web Services populates this field.

Note

What is idempotency?

When you make a mutating API request, the request typically returns a result before the operation's asynchronous workflows have completed. Operations might also time out or encounter other server issues before they complete, even though the request has already returned a result. This could make it difficult to determine whether the request succeeded or not, and could lead to multiple retries to ensure that the operation completes successfully. However, if the original request and the subsequent retries are successful, the operation is completed multiple times. This means that you might create more resources than you intended.

Idempotency ensures that an API request completes no more than one time. With an idempotent request, if the original request completes successfully, any subsequent retries complete successfully without performing any further actions.

"Tags": Tag you want to attach to the identity source.

Main.Mpa.delete_identity_source — Function

delete_identity_source(identity_source_arn)
delete_identity_source(identity_source_arn, params::Dict{String,<:Any})

Deletes an identity source. For more information, see Identity Source in the Multi-party approval User Guide.

Arguments

identity_source_arn: Amazon Resource Name (ARN) for identity source.

Main.Mpa.delete_inactive_approval_team_version — Function

delete_inactive_approval_team_version(arn, version_id)
delete_inactive_approval_team_version(arn, version_id, params::Dict{String,<:Any})

Deletes an inactive approval team. For more information, see Team health in the Multi-party approval User Guide.

You can also use this operation to delete a team draft. For more information, see Interacting with drafts in the Multi-party approval User Guide.

Arguments

arn: Amaazon Resource Name (ARN) for the team.

version_id: Version ID for the team.

Main.Mpa.get_approval_team — Function

get_approval_team(arn)
get_approval_team(arn, params::Dict{String,<:Any})

Returns details for an approval team.

Arguments

arn: Amazon Resource Name (ARN) for the team.

Main.Mpa.get_identity_source — Function

get_identity_source(identity_source_arn)
get_identity_source(identity_source_arn, params::Dict{String,<:Any})

Returns details for an identity source. For more information, see Identity Source in the Multi-party approval User Guide.

Arguments

identity_source_arn: Amazon Resource Name (ARN) for the identity source.

Main.Mpa.get_policy_version — Function

get_policy_version(policy_version_arn)
get_policy_version(policy_version_arn, params::Dict{String,<:Any})

Returns details for the version of a policy. Policies define the permissions for team resources.

Arguments

policy_version_arn: Amazon Resource Name (ARN) for the policy.

Main.Mpa.get_resource_policy — Function

get_resource_policy(policy_name, policy_type, resource_arn)
get_resource_policy(policy_name, policy_type, resource_arn, params::Dict{String,<:Any})

Returns details about a policy for a resource.

Arguments

policy_name: Name of the policy.

policy_type: The type of policy.

resource_arn: Amazon Resource Name (ARN) for the resource.

Main.Mpa.get_session — Function

get_session(session_arn)
get_session(session_arn, params::Dict{String,<:Any})

Returns details for an approval session. For more information, see Session in the Multi-party approval User Guide.

Arguments

session_arn: Amazon Resource Name (ARN) for the session.

Main.Mpa.list_approval_teams — Function

list_approval_teams()
list_approval_teams(params::Dict{String,<:Any})

Returns a list of approval teams.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of items to return in the response. If more results exist than the specified MaxResults value, a token is included in the response so that you can retrieve the remaining results.

"NextToken": If present, indicates that more output is available than is included in the current response. Use this value in the next_token request parameter in a next call to the operation to get more output. You can repeat this until the NextToken response element returns null.

Main.Mpa.list_identity_sources — Function

list_identity_sources()
list_identity_sources(params::Dict{String,<:Any})

Returns a list of identity sources. For more information, see Identity Source in the Multi-party approval User Guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of items to return in the response. If more results exist than the specified MaxResults value, a token is included in the response so that you can retrieve the remaining results.

"NextToken": If present, indicates that more output is available than is included in the current response. Use this value in the next_token request parameter in a next call to the operation to get more output. You can repeat this until the NextToken response element returns null.

Main.Mpa.list_policies — Function

list_policies()
list_policies(params::Dict{String,<:Any})

Returns a list of policies. Policies define the permissions for team resources.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of items to return in the response. If more results exist than the specified MaxResults value, a token is included in the response so that you can retrieve the remaining results.

"NextToken": If present, indicates that more output is available than is included in the current response. Use this value in the next_token request parameter in a next call to the operation to get more output. You can repeat this until the NextToken response element returns null.

Main.Mpa.list_policy_versions — Function

list_policy_versions(policy_arn)
list_policy_versions(policy_arn, params::Dict{String,<:Any})

Returns a list of the versions for policies. Policies define the permissions for team resources.

Arguments

policy_arn: Amazon Resource Name (ARN) for the policy.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of items to return in the response. If more results exist than the specified MaxResults value, a token is included in the response so that you can retrieve the remaining results.

"NextToken": If present, indicates that more output is available than is included in the current response. Use this value in the next_token request parameter in a next call to the operation to get more output. You can repeat this until the NextToken response element returns null.

Main.Mpa.list_resource_policies — Function

list_resource_policies(resource_arn)
list_resource_policies(resource_arn, params::Dict{String,<:Any})

Returns a list of policies for a resource.

Arguments

resource_arn: Amazon Resource Name (ARN) for the resource.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of items to return in the response. If more results exist than the specified MaxResults value, a token is included in the response so that you can retrieve the remaining results.

"NextToken": If present, indicates that more output is available than is included in the current response. Use this value in the next_token request parameter in a next call to the operation to get more output. You can repeat this until the NextToken response element returns null.

Main.Mpa.list_sessions — Function

list_sessions(approval_team_arn)
list_sessions(approval_team_arn, params::Dict{String,<:Any})

Returns a list of approval sessions. For more information, see Session in the Multi-party approval User Guide.

Arguments

approval_team_arn: Amazon Resource Name (ARN) for the approval team.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Filters": An array of Filter objects. Contains the filter to apply when listing sessions.

"MaxResults": The maximum number of items to return in the response. If more results exist than the specified MaxResults value, a token is included in the response so that you can retrieve the remaining results.

"NextToken": If present, indicates that more output is available than is included in the current response. Use this value in the next_token request parameter in a next call to the operation to get more output. You can repeat this until the NextToken response element returns null.

Main.Mpa.list_tags_for_resource — Function

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Returns a list of the tags for a resource.

Arguments

resource_arn: Amazon Resource Name (ARN) for the resource.

Main.Mpa.start_active_approval_team_deletion — Function

start_active_approval_team_deletion(arn)
start_active_approval_team_deletion(arn, params::Dict{String,<:Any})

Starts the deletion process for an active approval team.

Note

Deletions require team approval

Requests to delete an active team must be approved by the team.

Arguments

arn: Amazon Resource Name (ARN) for the team.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"PendingWindowDays": Number of days between when the team approves the delete request and when the team is deleted.

Main.Mpa.start_approval_team_baseline — Function

start_approval_team_baseline(arn)
start_approval_team_baseline(arn, params::Dict{String,<:Any})

Starts a baseline session for specified approvers on an ACTIVE approval team.

Arguments

arn: Amazon Resource Name (ARN) for the approval team.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ApproverIds": Array of approver IDs.

Main.Mpa.tag_resource — Function

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Creates or updates a resource tag. Each tag is a label consisting of a user-defined key and value. Tags can help you manage, identify, organize, search for, and filter resources.

Arguments

resource_arn: Amazon Resource Name (ARN) for the resource you want to tag.

tags: Tags that you have added to the specified resource.

Main.Mpa.untag_resource — Function

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes a resource tag. Each tag is a label consisting of a user-defined key and value. Tags can help you manage, identify, organize, search for, and filter resources.

Arguments

resource_arn: Amazon Resource Name (ARN) for the resource you want to untag.

tag_keys: Array of tag key-value pairs that you want to untag.

Main.Mpa.update_approval_team — Function

update_approval_team(arn)
update_approval_team(arn, params::Dict{String,<:Any})

Updates an approval team. You can request to update the team description, approval threshold, and approvers in the team.

Note

Updates require team approval

Updates to an active team must be approved by the team.

Arguments

arn: Amazon Resource Name (ARN) for the team.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ApprovalStrategy": An ApprovalStrategy object. Contains details for how the team grants approval.

"Approvers": An array of ApprovalTeamRequestApprover objects. Contains details for the approvers in the team.

"Description": Description for the team.

"UpdateActions": A list of UpdateAction to perform when updating the team.