Mwaa

This page documents function available when using the Mwaa module, created with @service Mwaa.

Index

Main.Mwaa.create_cli_token

Main.Mwaa.create_environment

Main.Mwaa.create_web_login_token

Main.Mwaa.delete_environment

Main.Mwaa.get_environment

Main.Mwaa.list_environments

Main.Mwaa.list_tags_for_resource

Main.Mwaa.publish_metrics

Main.Mwaa.tag_resource

Main.Mwaa.untag_resource

Main.Mwaa.update_environment

Documentation

Main.Mwaa.create_cli_token — Method

create_cli_token(name)
create_cli_token(name, params::Dict{String,<:Any})

Create a CLI token to use Airflow CLI.

Arguments

name: Create a CLI token request for a MWAA environment.

Main.Mwaa.create_environment — Method

create_environment(dag_s3_path, execution_role_arn, name, network_configuration, source_bucket_arn)
create_environment(dag_s3_path, execution_role_arn, name, network_configuration, source_bucket_arn, params::Dict{String,<:Any})

Creates an Amazon Managed Workflows for Apache Airflow (MWAA) environment.

Arguments

dag_s3_path: The relative path to the DAGs folder on your Amazon S3 bucket. For example, dags. To learn more, see Adding or updating DAGs.

execution_role_arn: The Amazon Resource Name (ARN) of the execution role for your environment. An execution role is an AWS Identity and Access Management (IAM) role that grants MWAA permission to access AWS services and resources used by your environment. For example, arn:aws:iam::123456789:role/my-execution-role. To learn more, see Amazon MWAA Execution role.

name: The name of the Amazon MWAA environment. For example, MyMWAAEnvironment.

network_configuration: The VPC networking components used to secure and enable network traffic between the AWS resources for your environment. To learn more, see About networking on Amazon MWAA.

source_bucket_arn: The Amazon Resource Name (ARN) of the Amazon S3 bucket where your DAG code and supporting files are stored. For example, arn:aws:s3:::my-airflow-bucket-unique-name. To learn more, see Create an Amazon S3 bucket for Amazon MWAA.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AirflowConfigurationOptions": A list of key-value pairs containing the Apache Airflow configuration options you want to attach to your environment. To learn more, see Apache Airflow configuration options.

"AirflowVersion": The Apache Airflow version for your environment. For example, v1.10.12. If no value is specified, defaults to the latest version. Valid values: v1.10.12.

"EnvironmentClass": The environment class type. Valid values: mw1.small, mw1.medium, mw1.large. To learn more, see Amazon MWAA environment class.

"KmsKey": The AWS Key Management Service (KMS) key to encrypt the data in your environment. You can use an AWS owned CMK, or a Customer managed CMK (advanced). To learn more, see Get started with Amazon Managed Workflows for Apache Airflow.

"LoggingConfiguration": Defines the Apache Airflow logs to send to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

"MaxWorkers": The maximum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. For example, 20. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the one worker that is included with your environment, or the number you specify in MinWorkers.

"MinWorkers": The minimum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the worker count you specify in the MinWorkers field. For example, 2.

"PluginsS3ObjectVersion": The version of the plugins.zip file on your Amazon S3 bucket. A version must be specified each time a plugins.zip file is updated. To learn more, see How S3 Versioning works.

"PluginsS3Path": The relative path to the plugins.zip file on your Amazon S3 bucket. For example, plugins.zip. If specified, then the plugins.zip version is required. To learn more, see Installing custom plugins.

"RequirementsS3ObjectVersion": The version of the requirements.txt file on your Amazon S3 bucket. A version must be specified each time a requirements.txt file is updated. To learn more, see How S3 Versioning works.

"RequirementsS3Path": The relative path to the requirements.txt file on your Amazon S3 bucket. For example, requirements.txt. If specified, then a file version is required. To learn more, see Installing Python dependencies.

"Schedulers": The number of Apache Airflow schedulers to run in your environment.

"Tags": The key-value tag pairs you want to associate to your environment. For example, "Environment": "Staging". To learn more, see Tagging AWS resources.

"WebserverAccessMode": The Apache Airflow Web server access mode. To learn more, see Apache Airflow access modes.

"WeeklyMaintenanceWindowStart": The day and time of the week to start weekly maintenance updates of your environment in the following format: DAY:HH:MM. For example: TUE:03:30. You can specify a start time in 30 minute increments only. Supported input includes the following: MON|TUE|WED|THU|FRI|SAT|SUN:([01]d|2[0-3]):(00|30)

Main.Mwaa.create_web_login_token — Method

create_web_login_token(name)
create_web_login_token(name, params::Dict{String,<:Any})

Create a JWT token to be used to login to Airflow Web UI with claims based Authentication.

Arguments

name: Create an Airflow Web UI login token request for a MWAA environment.

Main.Mwaa.delete_environment — Method

delete_environment(name)
delete_environment(name, params::Dict{String,<:Any})

Deletes an Amazon Managed Workflows for Apache Airflow (MWAA) environment.

Arguments

name: The name of the Amazon MWAA environment. For example, MyMWAAEnvironment.

Main.Mwaa.get_environment — Method

get_environment(name)
get_environment(name, params::Dict{String,<:Any})

Retrieves the details of an Amazon Managed Workflows for Apache Airflow (MWAA) environment.

Arguments

name: The name of the Amazon MWAA environment. For example, MyMWAAEnvironment.

Main.Mwaa.list_environments — Method

list_environments()
list_environments(params::Dict{String,<:Any})

Lists the Amazon Managed Workflows for Apache Airflow (MWAA) environments.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to retrieve per page. For example, 5 environments per page.

"NextToken": Retrieves the next page of the results.

Main.Mwaa.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists the key-value tag pairs associated to the Amazon Managed Workflows for Apache Airflow (MWAA) environment. For example, "Environment": "Staging".

Arguments

resource_arn: The Amazon Resource Name (ARN) of the Amazon MWAA environment. For example, arn:aws:airflow:us-east-1:123456789012:environment/MyMWAAEnvironment.

Main.Mwaa.publish_metrics — Method

publish_metrics(environment_name, metric_data)
publish_metrics(environment_name, metric_data, params::Dict{String,<:Any})

An operation for publishing metrics from the customers to the Ops plane.

Arguments

environment_name: Publishes environment metric data to Amazon CloudWatch.

metric_data: Publishes metric data points to Amazon CloudWatch. CloudWatch associates the data points with the specified metrica.

Main.Mwaa.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Associates key-value tag pairs to your Amazon Managed Workflows for Apache Airflow (MWAA) environment.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the Amazon MWAA environment. For example, arn:aws:airflow:us-east-1:123456789012:environment/MyMWAAEnvironment.

tags: The key-value tag pairs you want to associate to your environment. For example, "Environment": "Staging". To learn more, see Tagging AWS resources.

Main.Mwaa.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes key-value tag pairs associated to your Amazon Managed Workflows for Apache Airflow (MWAA) environment. For example, "Environment": "Staging".

Arguments

resource_arn: The Amazon Resource Name (ARN) of the Amazon MWAA environment. For example, arn:aws:airflow:us-east-1:123456789012:environment/MyMWAAEnvironment.

tag_keys: The key-value tag pair you want to remove. For example, "Environment": "Staging".

Main.Mwaa.update_environment — Method

update_environment(name)
update_environment(name, params::Dict{String,<:Any})

Updates an Amazon Managed Workflows for Apache Airflow (MWAA) environment.

Arguments

name: The name of your Amazon MWAA environment. For example, MyMWAAEnvironment.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AirflowConfigurationOptions": A list of key-value pairs containing the Apache Airflow configuration options you want to attach to your environment. To learn more, see Apache Airflow configuration options.

"AirflowVersion": The Apache Airflow version for your environment. For example, v1.10.12. If no value is specified, defaults to the latest version. Valid values: v1.10.12.

"DagS3Path": The relative path to the DAGs folder on your Amazon S3 bucket. For example, dags. To learn more, see Adding or updating DAGs.

"EnvironmentClass": The environment class type. Valid values: mw1.small, mw1.medium, mw1.large. To learn more, see Amazon MWAA environment class.

"ExecutionRoleArn": The Amazon Resource Name (ARN) of the execution role in IAM that allows MWAA to access AWS resources in your environment. For example, arn:aws:iam::123456789:role/my-execution-role. To learn more, see Amazon MWAA Execution role.

"LoggingConfiguration": Defines the Apache Airflow logs to send to CloudWatch Logs: DagProcessingLogs, SchedulerLogs, TaskLogs, WebserverLogs, WorkerLogs.

"MaxWorkers": The maximum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. For example, 20. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the one worker that is included with your environment, or the number you specify in MinWorkers.

"MinWorkers": The minimum number of workers that you want to run in your environment. MWAA scales the number of Apache Airflow workers up to the number you specify in the MaxWorkers field. When there are no more tasks running, and no more in the queue, MWAA disposes of the extra workers leaving the worker count you specify in the MinWorkers field. For example, 2.

"NetworkConfiguration": The VPC networking components used to secure and enable network traffic between the AWS resources for your environment. To learn more, see About networking on Amazon MWAA.

"PluginsS3ObjectVersion": The version of the plugins.zip file on your Amazon S3 bucket. A version must be specified each time a plugins.zip file is updated. To learn more, see How S3 Versioning works.

"PluginsS3Path": The relative path to the plugins.zip file on your Amazon S3 bucket. For example, plugins.zip. If specified, then the plugins.zip version is required. To learn more, see Installing custom plugins.

"RequirementsS3ObjectVersion": The version of the requirements.txt file on your Amazon S3 bucket. A version must be specified each time a requirements.txt file is updated. To learn more, see How S3 Versioning works.

"RequirementsS3Path": The relative path to the requirements.txt file on your Amazon S3 bucket. For example, requirements.txt. If specified, then a file version is required. To learn more, see Installing Python dependencies.

"Schedulers": The number of Apache Airflow schedulers to run in your Amazon MWAA environment.

"SourceBucketArn": The Amazon Resource Name (ARN) of the Amazon S3 bucket where your DAG code and supporting files are stored. For example, arn:aws:s3:::my-airflow-bucket-unique-name. To learn more, see Create an Amazon S3 bucket for Amazon MWAA.

"WebserverAccessMode": The Apache Airflow Web server access mode. To learn more, see Apache Airflow access modes.

"WeeklyMaintenanceWindowStart": The day and time of the week to start weekly maintenance updates of your environment in the following format: DAY:HH:MM. For example: TUE:03:30. You can specify a start time in 30 minute increments only. Supported input includes the following: MON|TUE|WED|THU|FRI|SAT|SUN:([01]d|2[0-3]):(00|30)