Neptune Graph

This page documents function available when using the Neptune_Graph module, created with @service Neptune_Graph.

Index

Main.Neptune_Graph.cancel_import_task

Main.Neptune_Graph.cancel_query

Main.Neptune_Graph.create_graph

Main.Neptune_Graph.create_graph_snapshot

Main.Neptune_Graph.create_graph_using_import_task

Main.Neptune_Graph.create_private_graph_endpoint

Main.Neptune_Graph.delete_graph

Main.Neptune_Graph.delete_graph_snapshot

Main.Neptune_Graph.delete_private_graph_endpoint

Main.Neptune_Graph.execute_query

Main.Neptune_Graph.get_graph

Main.Neptune_Graph.get_graph_snapshot

Main.Neptune_Graph.get_graph_summary

Main.Neptune_Graph.get_import_task

Main.Neptune_Graph.get_private_graph_endpoint

Main.Neptune_Graph.get_query

Main.Neptune_Graph.list_graph_snapshots

Main.Neptune_Graph.list_graphs

Main.Neptune_Graph.list_import_tasks

Main.Neptune_Graph.list_private_graph_endpoints

Main.Neptune_Graph.list_queries

Main.Neptune_Graph.list_tags_for_resource

Main.Neptune_Graph.reset_graph

Main.Neptune_Graph.restore_graph_from_snapshot

Main.Neptune_Graph.start_import_task

Main.Neptune_Graph.tag_resource

Main.Neptune_Graph.untag_resource

Main.Neptune_Graph.update_graph

Documentation

Main.Neptune_Graph.cancel_import_task — Function

cancel_import_task(task_identifier)
cancel_import_task(task_identifier, params::Dict{String,<:Any})

Deletes the specified import task.

Arguments

task_identifier: The unique identifier of the import task.

Main.Neptune_Graph.cancel_query — Function

cancel_query(graph_identifier, query_id)
cancel_query(graph_identifier, query_id, params::Dict{String,<:Any})

Cancels a specified query.

Arguments

graph_identifier: The unique identifier of the Neptune Analytics graph.

query_id: The unique identifier of the query to cancel.

Main.Neptune_Graph.create_graph — Function

create_graph(graph_name, provisioned_memory)
create_graph(graph_name, provisioned_memory, params::Dict{String,<:Any})

Creates a new Neptune Analytics graph.

Arguments

graph_name: A name for the new Neptune Analytics graph to be created.

The name must contain from 1 to 63 letters, numbers, or hyphens, and its first character must be a letter. It cannot end with a hyphen or contain two consecutive hyphens.

provisioned_memory: The provisioned memory-optimized Neptune Capacity Units (m-NCUs) to use for the graph. Min = 128

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"deletionProtection": Indicates whether or not to enable deletion protection on the graph. The graph can’t be deleted when deletion protection is enabled. (true or false).

"kmsKeyIdentifier": Specifies a KMS key to use to encrypt data in the new graph.

"publicConnectivity": Specifies whether or not the graph can be reachable over the internet. All access to graphs is IAM authenticated. (true to enable, or false to disable.

"replicaCount": The number of replicas in other AZs. Min =0, Max = 2, Default = 1.

Important

Additional charges equivalent to the m-NCUs selected for the graph apply for each replica.

"tags": Adds metadata tags to the new graph. These tags can also be used with cost allocation reporting, or used in a Condition statement in an IAM policy.

"vectorSearchConfiguration": Specifies the number of dimensions for vector embeddings that will be loaded into the graph. The value is specified as dimension=value. Max = 65,535

Main.Neptune_Graph.create_graph_snapshot — Function

create_graph_snapshot(graph_identifier, snapshot_name)
create_graph_snapshot(graph_identifier, snapshot_name, params::Dict{String,<:Any})

Creates a snapshot of the specific graph.

Arguments

graph_identifier: The unique identifier of the Neptune Analytics graph.

snapshot_name: The snapshot name. For example: my-snapshot-1.

The name must contain from 1 to 63 letters, numbers, or hyphens, and its first character must be a letter. It cannot end with a hyphen or contain two consecutive hyphens.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"tags": Adds metadata tags to the new graph. These tags can also be used with cost allocation reporting, or used in a Condition statement in an IAM policy.

Main.Neptune_Graph.create_graph_using_import_task — Function

create_graph_using_import_task(graph_name, role_arn, source)
create_graph_using_import_task(graph_name, role_arn, source, params::Dict{String,<:Any})

Creates a new Neptune Analytics graph and imports data into it, either from Amazon Simple Storage Service (S3) or from a Neptune database or a Neptune database snapshot.

The data can be loaded from files in S3 that in either the Gremlin CSV format or the openCypher load format.

Arguments

graph_name: A name for the new Neptune Analytics graph to be created.

The name must contain from 1 to 63 letters, numbers, or hyphens, and its first character must be a letter. It cannot end with a hyphen or contain two consecutive hyphens.

role_arn: The ARN of the IAM role that will allow access to the data that is to be imported.

source: A URL identifying to the location of the data to be imported. This can be an Amazon S3 path, or can point to a Neptune database endpoint or snapshot.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"deletionProtection": Indicates whether or not to enable deletion protection on the graph. The graph can’t be deleted when deletion protection is enabled. (true or false).

"failOnError": If set to true, the task halts when an import error is encountered. If set to false, the task skips the data that caused the error and continues if possible.

"format": Specifies the format of S3 data to be imported. Valid values are CSV, which identifies the Gremlin CSV format or OPENCYPHER, which identies the openCypher load format.

"importOptions": Contains options for controlling the import process. For example, if the failOnError key is set to false, the import skips problem data and attempts to continue (whereas if set to true, the default, or if omitted, the import operation halts immediately when an error is encountered.

"kmsKeyIdentifier": Specifies a KMS key to use to encrypt data imported into the new graph.

"maxProvisionedMemory": The maximum provisioned memory-optimized Neptune Capacity Units (m-NCUs) to use for the graph. Default: 1024, or the approved upper limit for your account.

If both the minimum and maximum values are specified, the max of the min-provisioned-memory and max-provisioned memory is used to create the graph. If neither value is specified 128 m-NCUs are used.

"minProvisionedMemory": The minimum provisioned memory-optimized Neptune Capacity Units (m-NCUs) to use for the graph. Default: 128

"publicConnectivity": Specifies whether or not the graph can be reachable over the internet. All access to graphs is IAM authenticated. (true to enable, or false to disable).

"replicaCount": The number of replicas in other AZs to provision on the new graph after import. Default = 0, Min = 0, Max = 2.

Important

Additional charges equivalent to the m-NCUs selected for the graph apply for each replica.

"tags": Adds metadata tags to the new graph. These tags can also be used with cost allocation reporting, or used in a Condition statement in an IAM policy.

"vectorSearchConfiguration": Specifies the number of dimensions for vector embeddings that will be loaded into the graph. The value is specified as dimension=value. Max = 65,535

Main.Neptune_Graph.create_private_graph_endpoint — Function

create_private_graph_endpoint(graph_identifier)
create_private_graph_endpoint(graph_identifier, params::Dict{String,<:Any})

Create a private graph endpoint to allow private access from to the graph from within a VPC. You can attach security groups to the private graph endpoint.

Note

VPC endpoint charges apply.

Arguments

graph_identifier: The unique identifier of the Neptune Analytics graph.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"subnetIds": Subnets in which private graph endpoint ENIs are created.

"vpcId": The VPC in which the private graph endpoint needs to be created.

"vpcSecurityGroupIds": Security groups to be attached to the private graph endpoint..

Main.Neptune_Graph.delete_graph — Function

delete_graph(graph_identifier, skip_snapshot)
delete_graph(graph_identifier, skip_snapshot, params::Dict{String,<:Any})

Deletes the specified graph. Graphs cannot be deleted if delete-protection is enabled.

Arguments

graph_identifier: The unique identifier of the Neptune Analytics graph.

skip_snapshot: Determines whether a final graph snapshot is created before the graph is deleted. If true is specified, no graph snapshot is created. If false is specified, a graph snapshot is created before the graph is deleted.

Main.Neptune_Graph.delete_graph_snapshot — Function

delete_graph_snapshot(snapshot_identifier)
delete_graph_snapshot(snapshot_identifier, params::Dict{String,<:Any})

Deletes the specifed graph snapshot.

Arguments

snapshot_identifier: ID of the graph snapshot to be deleted.

Main.Neptune_Graph.delete_private_graph_endpoint — Function

delete_private_graph_endpoint(graph_identifier, vpc_id)
delete_private_graph_endpoint(graph_identifier, vpc_id, params::Dict{String,<:Any})

Deletes a private graph endpoint.

Arguments

graph_identifier: The unique identifier of the Neptune Analytics graph.

vpc_id: The ID of the VPC where the private endpoint is located.

Main.Neptune_Graph.execute_query — Function

execute_query(graph_identifier, language, query_string)
execute_query(graph_identifier, language, query_string, params::Dict{String,<:Any})

Execute an openCypher query.

When invoking this operation in a Neptune Analytics cluster, the IAM user or role making the request must have a policy attached that allows one of the following IAM actions in that cluster, depending on the query:

neptune-graph:ReadDataViaQuery

neptune-graph:WriteDataViaQuery

neptune-graph:DeleteDataViaQuery

Arguments

graph_identifier: The unique identifier of the Neptune Analytics graph.

language: The query language the query is written in. Currently only openCypher is supported.

query_string: The query string to be executed.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"explainMode": The explain mode parameter returns a query explain instead of the actual query results. A query explain can be used to gather insights about the query execution such as planning decisions, time spent on each operator, solutions flowing etc.

"parameters": The data parameters the query can use in JSON format. For example: {"name": "john", "age": 20}. (optional)

"planCache": Query plan cache is a feature that saves the query plan and reuses it on successive executions of the same query. This reduces query latency, and works for both READ and UPDATE queries. The plan cache is an LRU cache with a 5 minute TTL and a capacity of 1000.

"queryTimeoutMilliseconds": Specifies the query timeout duration, in milliseconds. (optional)

Main.Neptune_Graph.get_graph — Function

get_graph(graph_identifier)
get_graph(graph_identifier, params::Dict{String,<:Any})

Gets information about a specified graph.

Arguments

graph_identifier: The unique identifier of the Neptune Analytics graph.

Main.Neptune_Graph.get_graph_snapshot — Function

get_graph_snapshot(snapshot_identifier)
get_graph_snapshot(snapshot_identifier, params::Dict{String,<:Any})

Retrieves a specified graph snapshot.

Arguments

snapshot_identifier: The ID of the snapshot to retrieve.

Main.Neptune_Graph.get_graph_summary — Function

get_graph_summary(graph_identifier)
get_graph_summary(graph_identifier, params::Dict{String,<:Any})

Gets a graph summary for a property graph.

Arguments

graph_identifier: The unique identifier of the Neptune Analytics graph.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"mode": The summary mode can take one of two values: basic (the default), and detailed.

Main.Neptune_Graph.get_import_task — Function

get_import_task(task_identifier)
get_import_task(task_identifier, params::Dict{String,<:Any})

Retrieves a specified import task.

Arguments

task_identifier: The unique identifier of the import task.

Main.Neptune_Graph.get_private_graph_endpoint — Function

get_private_graph_endpoint(graph_identifier, vpc_id)
get_private_graph_endpoint(graph_identifier, vpc_id, params::Dict{String,<:Any})

Retrieves information about a specified private endpoint.

Arguments

graph_identifier: The unique identifier of the Neptune Analytics graph.

vpc_id: The ID of the VPC where the private endpoint is located.

Main.Neptune_Graph.get_query — Function

get_query(graph_identifier, query_id)
get_query(graph_identifier, query_id, params::Dict{String,<:Any})

Retrieves the status of a specified query.

Note

When invoking this operation in a Neptune Analytics cluster, the IAM user or role making the request must have the neptune-graph:GetQueryStatus IAM action attached.

Arguments

graph_identifier: The unique identifier of the Neptune Analytics graph.

query_id: The ID of the query in question.

Main.Neptune_Graph.list_graph_snapshots — Function

list_graph_snapshots()
list_graph_snapshots(params::Dict{String,<:Any})

Lists available snapshots of a specified Neptune Analytics graph.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"graphIdentifier": The unique identifier of the Neptune Analytics graph.

"maxResults": The total number of records to return in the command's output.

If the total number of records available is more than the value specified, nextToken is provided in the command's output. To resume pagination, provide the nextToken output value in the nextToken argument of a subsequent command. Do not use the nextToken response element directly outside of the Amazon CLI.

"nextToken": Pagination token used to paginate output.

When this value is provided as input, the service returns results from where the previous response left off. When this value is present in output, it indicates that there are more results to retrieve.

Main.Neptune_Graph.list_graphs — Function

list_graphs()
list_graphs(params::Dict{String,<:Any})

Lists available Neptune Analytics graphs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The total number of records to return in the command's output.

If the total number of records available is more than the value specified, nextToken is provided in the command's output. To resume pagination, provide the nextToken output value in the nextToken argument of a subsequent command. Do not use the nextToken response element directly outside of the Amazon CLI.

"nextToken": Pagination token used to paginate output.

When this value is provided as input, the service returns results from where the previous response left off. When this value is present in output, it indicates that there are more results to retrieve.

Main.Neptune_Graph.list_import_tasks — Function

list_import_tasks()
list_import_tasks(params::Dict{String,<:Any})

Lists import tasks.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The total number of records to return in the command's output.

If the total number of records available is more than the value specified, nextToken is provided in the command's output. To resume pagination, provide the nextToken output value in the nextToken argument of a subsequent command. Do not use the nextToken response element directly outside of the Amazon CLI.

"nextToken": Pagination token used to paginate output.

When this value is provided as input, the service returns results from where the previous response left off. When this value is present in output, it indicates that there are more results to retrieve.

Main.Neptune_Graph.list_private_graph_endpoints — Function

list_private_graph_endpoints(graph_identifier)
list_private_graph_endpoints(graph_identifier, params::Dict{String,<:Any})

Lists private endpoints for a specified Neptune Analytics graph.

Arguments

graph_identifier: The unique identifier of the Neptune Analytics graph.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The total number of records to return in the command's output.

If the total number of records available is more than the value specified, nextToken is provided in the command's output. To resume pagination, provide the nextToken output value in the nextToken argument of a subsequent command. Do not use the nextToken response element directly outside of the Amazon CLI.

"nextToken": Pagination token used to paginate output.

When this value is provided as input, the service returns results from where the previous response left off. When this value is present in output, it indicates that there are more results to retrieve.

Main.Neptune_Graph.list_queries — Function

list_queries(graph_identifier, max_results)
list_queries(graph_identifier, max_results, params::Dict{String,<:Any})

Lists active openCypher queries.

Arguments

graph_identifier: The unique identifier of the Neptune Analytics graph.

max_results: The maximum number of results to be fetched by the API.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"state": Filtered list of queries based on state.

Main.Neptune_Graph.list_tags_for_resource — Function

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists tags associated with a specified resource.

Arguments

resource_arn: The ARN of the resource.

Main.Neptune_Graph.reset_graph — Function

reset_graph(graph_identifier, skip_snapshot)
reset_graph(graph_identifier, skip_snapshot, params::Dict{String,<:Any})

Empties the data from a specified Neptune Analytics graph.

Arguments

graph_identifier: ID of the graph to reset.

skip_snapshot: Determines whether a final graph snapshot is created before the graph data is deleted. If set to true, no graph snapshot is created. If set to false, a graph snapshot is created before the data is deleted.

Main.Neptune_Graph.restore_graph_from_snapshot — Function

restore_graph_from_snapshot(graph_name, snapshot_identifier)
restore_graph_from_snapshot(graph_name, snapshot_identifier, params::Dict{String,<:Any})

Restores a graph from a snapshot.

Arguments

graph_name: A name for the new Neptune Analytics graph to be created from the snapshot.

The name must contain from 1 to 63 letters, numbers, or hyphens, and its first character must be a letter. It cannot end with a hyphen or contain two consecutive hyphens.

snapshot_identifier: The ID of the snapshot in question.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"deletionProtection": A value that indicates whether the graph has deletion protection enabled. The graph can't be deleted when deletion protection is enabled.

"provisionedMemory": The provisioned memory-optimized Neptune Capacity Units (m-NCUs) to use for the graph.

Min = 128

"publicConnectivity": Specifies whether or not the graph can be reachable over the internet. All access to graphs is IAM authenticated. (true to enable, or false to disable).

"replicaCount": The number of replicas in other AZs. Min =0, Max = 2, Default =1

Important

Additional charges equivalent to the m-NCUs selected for the graph apply for each replica.

"tags": Adds metadata tags to the snapshot. These tags can also be used with cost allocation reporting, or used in a Condition statement in an IAM policy.

Main.Neptune_Graph.start_import_task — Function

start_import_task(graph_identifier, role_arn, source)
start_import_task(graph_identifier, role_arn, source, params::Dict{String,<:Any})

Import data into existing Neptune Analytics graph from Amazon Simple Storage Service (S3). The graph needs to be empty and in the AVAILABLE state.

Arguments

graph_identifier: The unique identifier of the Neptune Analytics graph.

role_arn: The ARN of the IAM role that will allow access to the data that is to be imported.

source: A URL identifying the location of the data to be imported. This can be an Amazon S3 path, or can point to a Neptune database endpoint or snapshot.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"failOnError": If set to true, the task halts when an import error is encountered. If set to false, the task skips the data that caused the error and continues if possible.

"format": Specifies the format of Amazon S3 data to be imported. Valid values are CSV, which identifies the Gremlin CSV format or OPENCYPHER, which identies the openCypher load format.

"importOptions":

Main.Neptune_Graph.tag_resource — Function

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Adds tags to the specified resource.

Arguments

resource_arn: ARN of the resource for which tags need to be added.

tags: The tags to be assigned to the Neptune Analytics resource.

The tags are metadata that are specified as a list of key-value pairs:

Key (string) – A key is the required name of the tag. The string value can be from 1 to 128 Unicode characters in length. It can't be prefixed with aws: and can only contain the set of Unicode characters specified by this Java regular expression: "^([\p{L}\p{Z}\p{N}_.:/=+\-]*)$").

Value (string) – A value is the optional value of the tag. The string value can be from 1 to 256 Unicode characters in length. It can't be prefixed with aws: and can only contain the set of Unicode characters specified by this Java regular expression: "^([\p{L}\p{Z}\p{N}_.:/=+\-]*)$").

Main.Neptune_Graph.untag_resource — Function

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes the specified tags from the specified resource.

Arguments

resource_arn: ARN of the resource whose tag needs to be removed.

tag_keys: Tag keys for the tags to be removed.

Main.Neptune_Graph.update_graph — Function

update_graph(graph_identifier)
update_graph(graph_identifier, params::Dict{String,<:Any})

Updates the configuration of a specified Neptune Analytics graph

Arguments

graph_identifier: The unique identifier of the Neptune Analytics graph.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"deletionProtection": A value that indicates whether the graph has deletion protection enabled. The graph can't be deleted when deletion protection is enabled.

"provisionedMemory": The provisioned memory-optimized Neptune Capacity Units (m-NCUs) to use for the graph.

Min = 128

"publicConnectivity": Specifies whether or not the graph can be reachable over the internet. All access to graphs is IAM authenticated. (true to enable, or false to disable.