Networkflowmonitor

This page documents function available when using the Networkflowmonitor module, created with @service Networkflowmonitor.

Index

Main.Networkflowmonitor.create_monitor

Main.Networkflowmonitor.create_scope

Main.Networkflowmonitor.delete_monitor

Main.Networkflowmonitor.delete_scope

Main.Networkflowmonitor.get_monitor

Main.Networkflowmonitor.get_query_results_monitor_top_contributors

Main.Networkflowmonitor.get_query_results_workload_insights_top_contributors

Main.Networkflowmonitor.get_query_results_workload_insights_top_contributors_data

Main.Networkflowmonitor.get_query_status_monitor_top_contributors

Main.Networkflowmonitor.get_query_status_workload_insights_top_contributors

Main.Networkflowmonitor.get_query_status_workload_insights_top_contributors_data

Main.Networkflowmonitor.get_scope

Main.Networkflowmonitor.list_monitors

Main.Networkflowmonitor.list_scopes

Main.Networkflowmonitor.list_tags_for_resource

Main.Networkflowmonitor.start_query_monitor_top_contributors

Main.Networkflowmonitor.start_query_workload_insights_top_contributors

Main.Networkflowmonitor.start_query_workload_insights_top_contributors_data

Main.Networkflowmonitor.stop_query_monitor_top_contributors

Main.Networkflowmonitor.stop_query_workload_insights_top_contributors

Main.Networkflowmonitor.stop_query_workload_insights_top_contributors_data

Main.Networkflowmonitor.tag_resource

Main.Networkflowmonitor.untag_resource

Main.Networkflowmonitor.update_monitor

Main.Networkflowmonitor.update_scope

Documentation

Main.Networkflowmonitor.create_monitor — Function

create_monitor(local_resources, monitor_name, scope_arn)
create_monitor(local_resources, monitor_name, scope_arn, params::Dict{String,<:Any})

Create a monitor for specific network flows between local and remote resources, so that you can monitor network performance for one or several of your workloads. For each monitor, Network Flow Monitor publishes detailed end-to-end performance metrics and a network health indicator (NHI) that informs you whether there were Amazon Web Services network issues for one or more of the network flows tracked by a monitor, during a time period that you choose.

Arguments

local_resources: The local resources to monitor. A local resource in a workload is the location of the host, or hosts, where the Network Flow Monitor agent is installed. For example, if a workload consists of an interaction between a web service and a backend database (for example, Amazon Dynamo DB), the subnet with the EC2 instance that hosts the web service, which also runs the agent, is the local resource.

Be aware that all local resources must belong to the current Region.

monitor_name: The name of the monitor.

scope_arn: The Amazon Resource Name (ARN) of the scope for the monitor.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive string of up to 64 ASCII characters that you specify to make an idempotent API request. Don't reuse the same client token for other API requests.

"remoteResources": The remote resources to monitor. A remote resource is the other endpoint in the bi-directional flow of a workload, with a local resource. For example, Amazon Dynamo DB can be a remote resource.

When you specify remote resources, be aware that specific combinations of resources are allowed and others are not, including the following constraints:

All remote resources that you specify must all belong to a single Region.

If you specify Amazon Web Services services as remote resources, any other remote resources that you specify must be in the current Region.

When you specify a remote resource for another Region, you can only specify the Region resource type. You cannot specify a subnet, VPC, or Availability Zone in another Region.

If you leave the RemoteResources parameter empty, the monitor will include all network flows that terminate in the current Region.

"tags": The tags for a monitor. You can add a maximum of 200 tags.

Main.Networkflowmonitor.create_scope — Function

create_scope(targets)
create_scope(targets, params::Dict{String,<:Any})

In Network Flow Monitor, you specify a scope for the service to generate metrics for. By using the scope, Network Flow Monitor can generate a topology of all the resources to measure performance metrics for. When you create a scope, you enable permissions for Network Flow Monitor.

A scope is a Region-account pair or multiple Region-account pairs. Network Flow Monitor uses your scope to determine all the resources (the topology) where Network Flow Monitor will gather network flow performance metrics for you. To provide performance metrics, Network Flow Monitor uses the data that is sent by the Network Flow Monitor agents you install on the resources.

To define the Region-account pairs for your scope, the Network Flow Monitor API uses the following constucts, which allow for future flexibility in defining scopes:

Targets, which are arrays of targetResources.

Target resources, which are Region-targetIdentifier pairs.

Target identifiers, made up of a targetID (currently always an account ID) and a targetType (currently always an account).

Arguments

targets: The targets to define the scope to be monitored. A target is an array of targetResources, which are currently Region-account pairs, defined by targetResource constructs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive string of up to 64 ASCII characters that you specify to make an idempotent API request. Don't reuse the same client token for other API requests.

"tags": The tags for a scope. You can add a maximum of 200 tags.

Main.Networkflowmonitor.delete_monitor — Function

delete_monitor(monitor_name)
delete_monitor(monitor_name, params::Dict{String,<:Any})

Deletes a monitor in Network Flow Monitor.

Arguments

monitor_name: The name of the monitor to delete.

Main.Networkflowmonitor.delete_scope — Function

delete_scope(scope_id)
delete_scope(scope_id, params::Dict{String,<:Any})

Deletes a scope that has been defined.

Arguments

scope_id: The identifier for the scope that includes the resources you want to get data results for. A scope ID is an internally-generated identifier that includes all the resources for a specific root account.

Main.Networkflowmonitor.get_monitor — Function

get_monitor(monitor_name)
get_monitor(monitor_name, params::Dict{String,<:Any})

Gets information about a monitor in Network Flow Monitor based on a monitor name. The information returned includes the Amazon Resource Name (ARN), create time, modified time, resources included in the monitor, and status information.

Arguments

monitor_name: The name of the monitor.

Main.Networkflowmonitor.get_query_results_monitor_top_contributors — Function

get_query_results_monitor_top_contributors(monitor_name, query_id)
get_query_results_monitor_top_contributors(monitor_name, query_id, params::Dict{String,<:Any})

Return the data for a query with the Network Flow Monitor query interface. You specify the query that you want to return results for by providing a query ID and a monitor name. This query returns the top contributors for a specific monitor.

Create a query ID for this call by calling the corresponding API call to start the query, StartQueryMonitorTopContributors. Use the scope ID that was returned for your account by CreateScope.

Top contributors in Network Flow Monitor are network flows with the highest values for a specific metric type. Top contributors can be across all workload insights, for a given scope, or for a specific monitor. Use the applicable call for the top contributors that you want to be returned.

Arguments

monitor_name: The name of the monitor.

query_id: The identifier for the query. A query ID is an internally-generated identifier for a specific query returned from an API call to create a query.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The number of query results that you want to return with this call.

"nextToken": The token for the next set of results. You receive this token from a previous call.

Main.Networkflowmonitor.get_query_results_workload_insights_top_contributors — Function

get_query_results_workload_insights_top_contributors(query_id, scope_id)
get_query_results_workload_insights_top_contributors(query_id, scope_id, params::Dict{String,<:Any})

Return the data for a query with the Network Flow Monitor query interface. You specify the query that you want to return results for by providing a query ID and a monitor name.

This query returns the top contributors for a scope for workload insights. Workload insights provide a high level view of network flow performance data collected by agents. To return the data for the top contributors, see GetQueryResultsWorkloadInsightsTopContributorsData.

Create a query ID for this call by calling the corresponding API call to start the query, StartQueryWorkloadInsightsTopContributors. Use the scope ID that was returned for your account by CreateScope.

Top contributors in Network Flow Monitor are network flows with the highest values for a specific metric type. Top contributors can be across all workload insights, for a given scope, or for a specific monitor. Use the applicable call for the top contributors that you want to be returned.

Arguments

query_id: The identifier for the query. A query ID is an internally-generated identifier for a specific query returned from an API call to create a query.

scope_id: The identifier for the scope that includes the resources you want to get data results for. A scope ID is an internally-generated identifier that includes all the resources for a specific root account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The number of query results that you want to return with this call.

"nextToken": The token for the next set of results. You receive this token from a previous call.

Main.Networkflowmonitor.get_query_results_workload_insights_top_contributors_data — Function

get_query_results_workload_insights_top_contributors_data(query_id, scope_id)
get_query_results_workload_insights_top_contributors_data(query_id, scope_id, params::Dict{String,<:Any})

Return the data for a query with the Network Flow Monitor query interface. Specify the query that you want to return results for by providing a query ID and a scope ID.

This query returns the data for top contributors for workload insights for a specific scope. Workload insights provide a high level view of network flow performance data collected by agents for a scope. To return just the top contributors, see GetQueryResultsWorkloadInsightsTopContributors.

Create a query ID for this call by calling the corresponding API call to start the query, StartQueryWorkloadInsightsTopContributorsData. Use the scope ID that was returned for your account by CreateScope.

Top contributors in Network Flow Monitor are network flows with the highest values for a specific metric type. Top contributors can be across all workload insights, for a given scope, or for a specific monitor. Use the applicable call for the top contributors that you want to be returned.

The top contributor network flows overall are for a specific metric type, for example, the number of retransmissions.

Arguments

query_id: The identifier for the query. A query ID is an internally-generated identifier for a specific query returned from an API call to create a query.

scope_id: The identifier for the scope that includes the resources you want to get data results for. A scope ID is an internally-generated identifier that includes all the resources for a specific root account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The number of query results that you want to return with this call.

"nextToken": The token for the next set of results. You receive this token from a previous call.

Main.Networkflowmonitor.get_query_status_monitor_top_contributors — Function

get_query_status_monitor_top_contributors(monitor_name, query_id)
get_query_status_monitor_top_contributors(monitor_name, query_id, params::Dict{String,<:Any})

Returns the current status of a query for the Network Flow Monitor query interface, for a specified query ID and monitor. This call returns the query status for the top contributors for a monitor.

When you create a query, use this call to check the status of the query to make sure that it has has SUCCEEDED before you review the results. Use the same query ID that you used for the corresponding API call to start (create) the query, StartQueryMonitorTopContributors.

When you run a query, use this call to check the status of the query to make sure that the query has SUCCEEDED before you review the results.

Arguments

monitor_name: The name of the monitor.

query_id: The identifier for the query. A query ID is an internally-generated identifier for a specific query returned from an API call to start a query.

Main.Networkflowmonitor.get_query_status_workload_insights_top_contributors — Function

get_query_status_workload_insights_top_contributors(query_id, scope_id)
get_query_status_workload_insights_top_contributors(query_id, scope_id, params::Dict{String,<:Any})

Return the data for a query with the Network Flow Monitor query interface. Specify the query that you want to return results for by providing a query ID and a monitor name. This query returns the top contributors for workload insights.

When you start a query, use this call to check the status of the query to make sure that it has has SUCCEEDED before you review the results. Use the same query ID that you used for the corresponding API call to start the query, StartQueryWorkloadInsightsTopContributors.

Top contributors in Network Flow Monitor are network flows with the highest values for a specific metric type. Top contributors can be across all workload insights, for a given scope, or for a specific monitor. Use the applicable call for the top contributors that you want to be returned.

Arguments

query_id: The identifier for the query. A query ID is an internally-generated identifier for a specific query returned from an API call to start a query.

scope_id: The identifier for the scope that includes the resources you want to get data results for. A scope ID is an internally-generated identifier that includes all the resources for a specific root account.

Main.Networkflowmonitor.get_query_status_workload_insights_top_contributors_data — Function

get_query_status_workload_insights_top_contributors_data(query_id, scope_id)
get_query_status_workload_insights_top_contributors_data(query_id, scope_id, params::Dict{String,<:Any})

Returns the current status of a query for the Network Flow Monitor query interface, for a specified query ID and monitor. This call returns the query status for the top contributors data for workload insights.

When you start a query, use this call to check the status of the query to make sure that it has has SUCCEEDED before you review the results. Use the same query ID that you used for the corresponding API call to start the query, StartQueryWorkloadInsightsTopContributorsData.

Top contributors in Network Flow Monitor are network flows with the highest values for a specific metric type. Top contributors can be across all workload insights, for a given scope, or for a specific monitor. Use the applicable call for the top contributors that you want to be returned.

The top contributor network flows overall are for a specific metric type, for example, the number of retransmissions.

Arguments

query_id: The identifier for the query. A query ID is an internally-generated identifier for a specific query returned from an API call to start a query.

scope_id: The identifier for the scope that includes the resources you want to get data results for. A scope ID is an internally-generated identifier that includes all the resources for a specific root account. A scope ID is returned from a CreateScope API call.

Main.Networkflowmonitor.get_scope — Function

get_scope(scope_id)
get_scope(scope_id, params::Dict{String,<:Any})

Gets information about a scope, including the name, status, tags, and target details. The scope in Network Flow Monitor is an account.

Arguments

scope_id: The identifier for the scope that includes the resources you want to get data results for. A scope ID is an internally-generated identifier that includes all the resources for a specific root account. A scope ID is returned from a CreateScope API call.

Main.Networkflowmonitor.list_monitors — Function

list_monitors()
list_monitors(params::Dict{String,<:Any})

List all monitors in an account. Optionally, you can list only monitors that have a specific status, by using the STATUS parameter.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The number of query results that you want to return with this call.

"monitorStatus": The status of a monitor. The status can be one of the following

PENDING: The monitor is in the process of being created.

ACTIVE: The monitor is active.

INACTIVE: The monitor is inactive.

ERROR: Monitor creation failed due to an error.

DELETING: The monitor is in the process of being deleted.

"nextToken": The token for the next set of results. You receive this token from a previous call.

Main.Networkflowmonitor.list_scopes — Function

list_scopes()
list_scopes(params::Dict{String,<:Any})

List all the scopes for an account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The number of query results that you want to return with this call.

"nextToken": The token for the next set of results. You receive this token from a previous call.

Main.Networkflowmonitor.list_tags_for_resource — Function

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Returns all the tags for a resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

Main.Networkflowmonitor.start_query_monitor_top_contributors — Function

start_query_monitor_top_contributors(destination_category, end_time, metric_name, monitor_name, start_time)
start_query_monitor_top_contributors(destination_category, end_time, metric_name, monitor_name, start_time, params::Dict{String,<:Any})

Create a query that you can use with the Network Flow Monitor query interface to return the top contributors for a monitor. Specify the monitor that you want to create the query for.

The call returns a query ID that you can use with GetQueryResultsMonitorTopContributors to run the query and return the top contributors for a specific monitor.

Top contributors in Network Flow Monitor are network flows with the highest values for a specific metric type. Top contributors can be across all workload insights, for a given scope, or for a specific monitor. Use the applicable APIs for the top contributors that you want to be returned.

Arguments

destination_category: The category that you want to query top contributors for, for a specific monitor. Destination categories can be one of the following:

INTRA_AZ: Top contributor network flows within a single Availability Zone

INTER_AZ: Top contributor network flows between Availability Zones

INTER_REGION: Top contributor network flows between Regions (to the edge of another Region)

INTER_VPC: Top contributor network flows between VPCs

AMAZON_S3: Top contributor network flows to or from Amazon S3

AMAZON_DYNAMODB: Top contributor network flows to or from Amazon Dynamo DB

UNCLASSIFIED: Top contributor network flows that do not have a bucket classification

end_time: The timestamp that is the date and time end of the period that you want to retrieve results for with your query.

metric_name: The metric that you want to query top contributors for. That is, you can specify a metric with this call and return the top contributor network flows, for that type of metric, for a monitor and (optionally) within a specific category, such as network flows between Availability Zones.

monitor_name: The name of the monitor.

start_time: The timestamp that is the date and time that is the beginning of the period that you want to retrieve results for with your query.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"limit": The maximum number of top contributors to return.

Main.Networkflowmonitor.start_query_workload_insights_top_contributors — Function

start_query_workload_insights_top_contributors(destination_category, end_time, metric_name, scope_id, start_time)
start_query_workload_insights_top_contributors(destination_category, end_time, metric_name, scope_id, start_time, params::Dict{String,<:Any})

Create a query with the Network Flow Monitor query interface that you can run to return workload insights top contributors. Specify the scope that you want to create a query for.

The call returns a query ID that you can use with GetQueryResultsWorkloadInsightsTopContributors to run the query and return the top contributors for the workload insights for a scope.

Top contributors in Network Flow Monitor are network flows with the highest values for a specific metric type. Top contributors can be across all workload insights, for a given scope, or for a specific monitor. Use the applicable APIs for the top contributors that you want to be returned.

Arguments

destination_category: The destination category for a top contributors row. Destination categories can be one of the following:

INTRA_AZ: Top contributor network flows within a single Availability Zone

INTER_AZ: Top contributor network flows between Availability Zones

INTER_REGION: Top contributor network flows between Regions (to the edge of another Region)

INTER_VPC: Top contributor network flows between VPCs

AWS_SERVICES: Top contributor network flows to or from Amazon Web Services services

UNCLASSIFIED: Top contributor network flows that do not have a bucket classification

end_time: The timestamp that is the date and time end of the period that you want to retrieve results for with your query.

metric_name: The metric that you want to query top contributors for. That is, you can specify this metric to return the top contributor network flows, for this type of metric, for a monitor and (optionally) within a specific category, such as network flows between Availability Zones.

scope_id: The identifier for the scope that includes the resources you want to get data results for. A scope ID is an internally-generated identifier that includes all the resources for a specific root account. A scope ID is returned from a CreateScope API call.

start_time: The timestamp that is the date and time that is the beginning of the period that you want to retrieve results for with your query.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"limit": The maximum number of top contributors to return.

Main.Networkflowmonitor.start_query_workload_insights_top_contributors_data — Function

start_query_workload_insights_top_contributors_data(destination_category, end_time, metric_name, scope_id, start_time)
start_query_workload_insights_top_contributors_data(destination_category, end_time, metric_name, scope_id, start_time, params::Dict{String,<:Any})

Create a query with the Network Flow Monitor query interface that you can run to return data for workload insights top contributors. Specify the scope that you want to create a query for.

The call returns a query ID that you can use with GetQueryResultsWorkloadInsightsTopContributorsData to run the query and return the data for the top contributors for the workload insights for a scope.

Top contributors in Network Flow Monitor are network flows with the highest values for a specific metric type. Top contributors can be across all workload insights, for a given scope, or for a specific monitor. Use the applicable call for the top contributors that you want to be returned.

Arguments

destination_category: The destination category for a top contributors. Destination categories can be one of the following:

INTRA_AZ: Top contributor network flows within a single Availability Zone

INTER_AZ: Top contributor network flows between Availability Zones

INTER_REGION: Top contributor network flows between Regions (to the edge of another Region)

INTER_VPC: Top contributor network flows between VPCs

AWS_SERVICES: Top contributor network flows to or from Amazon Web Services services

UNCLASSIFIED: Top contributor network flows that do not have a bucket classification

end_time: The timestamp that is the date and time end of the period that you want to retrieve results for with your query.

metric_name: The metric that you want to query top contributors for. That is, you can specify this metric to return the top contributor network flows, for this type of metric, for a monitor and (optionally) within a specific category, such as network flows between Availability Zones.

scope_id: The identifier for the scope that includes the resources you want to get data results for. A scope ID is an internally-generated identifier that includes all the resources for a specific root account.

start_time: The timestamp that is the date and time that is the beginning of the period that you want to retrieve results for with your query.

Main.Networkflowmonitor.stop_query_monitor_top_contributors — Function

stop_query_monitor_top_contributors(monitor_name, query_id)
stop_query_monitor_top_contributors(monitor_name, query_id, params::Dict{String,<:Any})

Stop a top contributors query for a monitor. Specify the query that you want to stop by providing a query ID and a monitor name.

Top contributors in Network Flow Monitor are network flows with the highest values for a specific metric type. Top contributors can be across all workload insights, for a given scope, or for a specific monitor. Use the applicable call for the top contributors that you want to be returned.

Arguments

monitor_name: The name of the monitor.

query_id: The identifier for the query. A query ID is an internally-generated identifier for a specific query returned from an API call to create a query.

Main.Networkflowmonitor.stop_query_workload_insights_top_contributors — Function

stop_query_workload_insights_top_contributors(query_id, scope_id)
stop_query_workload_insights_top_contributors(query_id, scope_id, params::Dict{String,<:Any})

Stop a top contributors query for workload insights. Specify the query that you want to stop by providing a query ID and a scope ID.

Top contributors in Network Flow Monitor are network flows with the highest values for a specific metric type. Top contributors can be across all workload insights, for a given scope, or for a specific monitor. Use the applicable call for the top contributors that you want to be returned.

Arguments

query_id: The identifier for the query. A query ID is an internally-generated identifier for a specific query returned from an API call to create a query.

scope_id: The identifier for the scope that includes the resources you want to get data results for. A scope ID is an internally-generated identifier that includes all the resources for a specific root account.

Main.Networkflowmonitor.stop_query_workload_insights_top_contributors_data — Function

stop_query_workload_insights_top_contributors_data(query_id, scope_id)
stop_query_workload_insights_top_contributors_data(query_id, scope_id, params::Dict{String,<:Any})

Stop a top contributors data query for workload insights. Specify the query that you want to stop by providing a query ID and a scope ID.

Top contributors in Network Flow Monitor are network flows with the highest values for a specific metric type. Top contributors can be across all workload insights, for a given scope, or for a specific monitor. Use the applicable call for the top contributors that you want to be returned.

Arguments

query_id: The identifier for the query. A query ID is an internally-generated identifier for a specific query returned from an API call to create a query.

scope_id: The identifier for the scope that includes the resources you want to get data results for. A scope ID is an internally-generated identifier that includes all the resources for a specific root account.

Main.Networkflowmonitor.tag_resource — Function

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Adds a tag to a resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

tags: The tags for a resource.

Main.Networkflowmonitor.untag_resource — Function

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes a tag from a resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

tag_keys: Keys that you specified when you tagged a resource.

Main.Networkflowmonitor.update_monitor — Function

update_monitor(monitor_name)
update_monitor(monitor_name, params::Dict{String,<:Any})

Update a monitor to add or remove local or remote resources.

Arguments

monitor_name: The name of the monitor.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive string of up to 64 ASCII characters that you specify to make an idempotent API request. Don't reuse the same client token for other API requests.

"localResourcesToAdd": Additional local resources to specify network flows for a monitor, as an array of resources with identifiers and types. A local resource in a workload is the location of hosts where the Network Flow Monitor agent is installed.

"localResourcesToRemove": The local resources to remove, as an array of resources with identifiers and types.

"remoteResourcesToAdd": The remote resources to add, as an array of resources with identifiers and types.

A remote resource is the other endpoint in the flow of a workload, with a local resource. For example, Amazon Dynamo DB can be a remote resource.

"remoteResourcesToRemove": The remote resources to remove, as an array of resources with identifiers and types.

A remote resource is the other endpoint specified for the network flow of a workload, with a local resource. For example, Amazon Dynamo DB can be a remote resource.

Main.Networkflowmonitor.update_scope — Function

update_scope(scope_id)
update_scope(scope_id, params::Dict{String,<:Any})

Update a scope to add or remove resources that you want to be available for Network Flow Monitor to generate metrics for, when you have active agents on those resources sending metrics reports to the Network Flow Monitor backend.

Arguments

scope_id: The identifier for the scope that includes the resources you want to get data results for. A scope ID is an internally-generated identifier that includes all the resources for a specific root account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"resourcesToAdd": A list of resources to add to a scope.

"resourcesToDelete": A list of resources to delete from a scope.