Nimble

This page documents function available when using the Nimble module, created with @service Nimble.

Index

Main.Nimble.accept_eulas

Main.Nimble.create_launch_profile

Main.Nimble.create_streaming_image

Main.Nimble.create_streaming_session

Main.Nimble.create_streaming_session_stream

Main.Nimble.create_studio

Main.Nimble.create_studio_component

Main.Nimble.delete_launch_profile

Main.Nimble.delete_launch_profile_member

Main.Nimble.delete_streaming_image

Main.Nimble.delete_streaming_session

Main.Nimble.delete_studio

Main.Nimble.delete_studio_component

Main.Nimble.delete_studio_member

Main.Nimble.get_eula

Main.Nimble.get_launch_profile

Main.Nimble.get_launch_profile_details

Main.Nimble.get_launch_profile_initialization

Main.Nimble.get_launch_profile_member

Main.Nimble.get_streaming_image

Main.Nimble.get_streaming_session

Main.Nimble.get_streaming_session_stream

Main.Nimble.get_studio

Main.Nimble.get_studio_component

Main.Nimble.get_studio_member

Main.Nimble.list_eula_acceptances

Main.Nimble.list_eulas

Main.Nimble.list_launch_profile_members

Main.Nimble.list_launch_profiles

Main.Nimble.list_streaming_images

Main.Nimble.list_streaming_sessions

Main.Nimble.list_studio_components

Main.Nimble.list_studio_members

Main.Nimble.list_studios

Main.Nimble.list_tags_for_resource

Main.Nimble.put_launch_profile_members

Main.Nimble.put_studio_members

Main.Nimble.start_studio_ssoconfiguration_repair

Main.Nimble.tag_resource

Main.Nimble.untag_resource

Main.Nimble.update_launch_profile

Main.Nimble.update_launch_profile_member

Main.Nimble.update_streaming_image

Main.Nimble.update_studio

Main.Nimble.update_studio_component

Documentation

Main.Nimble.accept_eulas — Method

accept_eulas(studio_id)
accept_eulas(studio_id, params::Dict{String,<:Any})

Accept EULAs.

Arguments

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"X-Amz-Client-Token": To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

"eulaIds": The EULA ID.

Main.Nimble.create_launch_profile — Method

create_launch_profile(ec2_subnet_ids, launch_profile_protocol_versions, name, stream_configuration, studio_component_ids, studio_id)
create_launch_profile(ec2_subnet_ids, launch_profile_protocol_versions, name, stream_configuration, studio_component_ids, studio_id, params::Dict{String,<:Any})

Create a launch profile.

Arguments

ec2_subnet_ids:

launch_profile_protocol_versions: The version number of the protocol that is used by the launch profile. The only valid version is "2021-03-31".

name: The name for the launch profile.

stream_configuration: A configuration for a streaming session.

studio_component_ids: Unique identifiers for a collection of studio components that can be used with this launch profile.

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"X-Amz-Client-Token": To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

"description": The description.

"tags": A collection of labels, in the form of key:value pairs, that apply to this resource.

Main.Nimble.create_streaming_image — Method

create_streaming_image(ec2_image_id, name, studio_id)
create_streaming_image(ec2_image_id, name, studio_id, params::Dict{String,<:Any})

Creates a streaming image resource in a studio.

Arguments

ec2_image_id: The ID of an EC2 machine image with which to create this streaming image.

name: A friendly name for a streaming image resource.

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"X-Amz-Client-Token": To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

"description": A human-readable description of the streaming image.

"tags": A collection of labels, in the form of key:value pairs, that apply to this resource.

Main.Nimble.create_streaming_session — Method

create_streaming_session(studio_id)
create_streaming_session(studio_id, params::Dict{String,<:Any})

Creates a streaming session in a studio. After invoking this operation, you must poll GetStreamingSession until the streaming session is in state READY.

Arguments

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"X-Amz-Client-Token": To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

"ec2InstanceType": The EC2 Instance type used for the streaming session.

"launchProfileId": The launch profile ID.

"ownedBy": The user ID of the user that owns the streaming session.

"streamingImageId": The ID of the streaming image.

"tags": A collection of labels, in the form of key:value pairs, that apply to this resource.

Main.Nimble.create_streaming_session_stream — Method

create_streaming_session_stream(session_id, studio_id)
create_streaming_session_stream(session_id, studio_id, params::Dict{String,<:Any})

Creates a streaming session stream for a streaming session. After invoking this API, invoke GetStreamingSessionStream with the returned streamId to poll the resource until it is in state READY.

Arguments

session_id: The session ID.

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"X-Amz-Client-Token": To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

"expirationInSeconds": The expiration time in seconds.

Main.Nimble.create_studio — Method

create_studio(admin_role_arn, display_name, studio_name, user_role_arn)
create_studio(admin_role_arn, display_name, studio_name, user_role_arn, params::Dict{String,<:Any})

Create a new Studio. When creating a Studio, two IAM roles must be provided: the admin role and the user Role. These roles are assumed by your users when they log in to the Nimble Studio portal. The user role must have the AmazonNimbleStudio-StudioUser managed policy attached for the portal to function properly. The Admin Role must have the AmazonNimbleStudio-StudioAdmin managed policy attached for the portal to function properly. You may optionally specify a KMS key in the StudioEncryptionConfiguration. In Nimble Studio, resource names, descriptions, initialization scripts, and other data you provide are always encrypted at rest using an KMS key. By default, this key is owned by Amazon Web Services and managed on your behalf. You may provide your own KMS key when calling CreateStudio to encrypt this data using a key you own and manage. When providing an KMS key during studio creation, Nimble Studio creates KMS grants in your account to provide your studio user and admin roles access to these KMS keys. If you delete this grant, the studio will no longer be accessible to your portal users. If you delete the studio KMS key, your studio will no longer be accessible.

Arguments

admin_role_arn: The IAM role that Studio Admins will assume when logging in to the Nimble Studio portal.

display_name: A friendly name for the studio.

studio_name: The studio name that is used in the URL of the Nimble Studio portal when accessed by Nimble Studio users.

user_role_arn: The IAM role that Studio Users will assume when logging in to the Nimble Studio portal.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"X-Amz-Client-Token": To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

"studioEncryptionConfiguration": The studio encryption configuration.

"tags": A collection of labels, in the form of key:value pairs, that apply to this resource.

Main.Nimble.create_studio_component — Method

create_studio_component(name, studio_id, type)
create_studio_component(name, studio_id, type, params::Dict{String,<:Any})

Creates a studio component resource.

Arguments

name: The name for the studio component.

studio_id: The studio ID.

type: The type of the studio component.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"X-Amz-Client-Token": To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

"configuration": The configuration of the studio component, based on component type.

"description": The description.

"ec2SecurityGroupIds": The EC2 security groups that control access to the studio component.

"initializationScripts": Initialization scripts for studio components.

"scriptParameters": Parameters for the studio component scripts.

"subtype": The specific subtype of a studio component.

"tags": A collection of labels, in the form of key:value pairs, that apply to this resource.

Main.Nimble.delete_launch_profile — Method

delete_launch_profile(launch_profile_id, studio_id)
delete_launch_profile(launch_profile_id, studio_id, params::Dict{String,<:Any})

Permanently delete a launch profile.

Arguments

launch_profile_id: The launch profile ID.

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"X-Amz-Client-Token": To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

Main.Nimble.delete_launch_profile_member — Method

delete_launch_profile_member(launch_profile_id, principal_id, studio_id)
delete_launch_profile_member(launch_profile_id, principal_id, studio_id, params::Dict{String,<:Any})

Delete a user from launch profile membership.

Arguments

launch_profile_id: The launch profile ID.

principal_id: The principal ID.

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"X-Amz-Client-Token": To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

Main.Nimble.delete_streaming_image — Method

delete_streaming_image(streaming_image_id, studio_id)
delete_streaming_image(streaming_image_id, studio_id, params::Dict{String,<:Any})

Delete streaming image.

Arguments

streaming_image_id: The streaming image ID.

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"X-Amz-Client-Token": To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

Main.Nimble.delete_streaming_session — Method

delete_streaming_session(session_id, studio_id)
delete_streaming_session(session_id, studio_id, params::Dict{String,<:Any})

Deletes streaming session resource. After invoking this operation, use GetStreamingSession to poll the resource until it transitions to a DELETED state. A streaming session will count against your streaming session quota until it is marked DELETED.

Arguments

session_id: The session ID.

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"X-Amz-Client-Token": To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

Main.Nimble.delete_studio — Method

delete_studio(studio_id)
delete_studio(studio_id, params::Dict{String,<:Any})

Delete a studio resource.

Arguments

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"X-Amz-Client-Token": To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

Main.Nimble.delete_studio_component — Method

delete_studio_component(studio_component_id, studio_id)
delete_studio_component(studio_component_id, studio_id, params::Dict{String,<:Any})

Deletes a studio component resource.

Arguments

studio_component_id: The studio component ID.

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"X-Amz-Client-Token": To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

Main.Nimble.delete_studio_member — Method

delete_studio_member(principal_id, studio_id)
delete_studio_member(principal_id, studio_id, params::Dict{String,<:Any})

Delete a user from studio membership.

Arguments

principal_id: The principal ID.

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"X-Amz-Client-Token": To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

Main.Nimble.get_eula — Method

get_eula(eula_id)
get_eula(eula_id, params::Dict{String,<:Any})

Get Eula.

Arguments

eula_id: The EULA ID.

Main.Nimble.get_launch_profile — Method

get_launch_profile(launch_profile_id, studio_id)
get_launch_profile(launch_profile_id, studio_id, params::Dict{String,<:Any})

Get a launch profile.

Arguments

launch_profile_id: The launch profile ID.

studio_id: The studio ID.

Main.Nimble.get_launch_profile_details — Method

get_launch_profile_details(launch_profile_id, studio_id)
get_launch_profile_details(launch_profile_id, studio_id, params::Dict{String,<:Any})

Launch profile details include the launch profile resource and summary information of resources that are used by, or available to, the launch profile. This includes the name and description of all studio components used by the launch profiles, and the name and description of streaming images that can be used with this launch profile.

Arguments

launch_profile_id: The launch profile ID.

studio_id: The studio ID.

Main.Nimble.get_launch_profile_initialization — Method

get_launch_profile_initialization(launch_profile_id, launch_profile_protocol_versions, launch_purpose, platform, studio_id)
get_launch_profile_initialization(launch_profile_id, launch_profile_protocol_versions, launch_purpose, platform, studio_id, params::Dict{String,<:Any})

Get a launch profile initialization.

Arguments

launch_profile_id: The launch profile ID.

launch_profile_protocol_versions: A collection of launch profile protocol versions.

launch_purpose: The launch purpose.

platform: The platform.

studio_id: The studio ID.

Main.Nimble.get_launch_profile_member — Method

get_launch_profile_member(launch_profile_id, principal_id, studio_id)
get_launch_profile_member(launch_profile_id, principal_id, studio_id, params::Dict{String,<:Any})

Get a user persona in launch profile membership.

Arguments

launch_profile_id: The launch profile ID.

principal_id: The principal ID.

studio_id: The studio ID.

Main.Nimble.get_streaming_image — Method

get_streaming_image(streaming_image_id, studio_id)
get_streaming_image(streaming_image_id, studio_id, params::Dict{String,<:Any})

Get streaming image.

Arguments

streaming_image_id: The streaming image ID.

studio_id: The studio ID.

Main.Nimble.get_streaming_session — Method

get_streaming_session(session_id, studio_id)
get_streaming_session(session_id, studio_id, params::Dict{String,<:Any})

Gets StreamingSession resource. Invoke this operation to poll for a streaming session state while creating or deleting a session.

Arguments

session_id: The session ID.

studio_id: The studio ID.

Main.Nimble.get_streaming_session_stream — Method

get_streaming_session_stream(session_id, stream_id, studio_id)
get_streaming_session_stream(session_id, stream_id, studio_id, params::Dict{String,<:Any})

Gets a StreamingSessionStream for a streaming session. Invoke this operation to poll the resource after invoking CreateStreamingSessionStream. After the StreamingSessionStream changes to the state READY, the url property will contain a stream to be used with the DCV streaming client.

Arguments

session_id: The session ID.

stream_id: The stream ID.

studio_id: The studio ID.

Main.Nimble.get_studio — Method

get_studio(studio_id)
get_studio(studio_id, params::Dict{String,<:Any})

Get a Studio resource.

Arguments

studio_id: The studio ID.

Main.Nimble.get_studio_component — Method

get_studio_component(studio_component_id, studio_id)
get_studio_component(studio_component_id, studio_id, params::Dict{String,<:Any})

Gets a studio component resource.

Arguments

studio_component_id: The studio component ID.

studio_id: The studio ID.

Main.Nimble.get_studio_member — Method

get_studio_member(principal_id, studio_id)
get_studio_member(principal_id, studio_id, params::Dict{String,<:Any})

Get a user's membership in a studio.

Arguments

principal_id: The principal ID.

studio_id: The studio ID.

Main.Nimble.list_eula_acceptances — Method

list_eula_acceptances(studio_id)
list_eula_acceptances(studio_id, params::Dict{String,<:Any})

List Eula Acceptances.

Arguments

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"eulaIds": A collection of EULA IDs.

"nextToken": The token for the next set of results, or null if there are no more results.

Main.Nimble.list_eulas — Method

list_eulas()
list_eulas(params::Dict{String,<:Any})

List Eulas.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"eulaIds": A collection of EULA IDs.

"nextToken": The token for the next set of results, or null if there are no more results.

Main.Nimble.list_launch_profile_members — Method

list_launch_profile_members(launch_profile_id, studio_id)
list_launch_profile_members(launch_profile_id, studio_id, params::Dict{String,<:Any})

Get all users in a given launch profile membership.

Arguments

launch_profile_id: The launch profile ID.

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to be returned per request.

"nextToken": The token for the next set of results, or null if there are no more results.

Main.Nimble.list_launch_profiles — Method

list_launch_profiles(studio_id)
list_launch_profiles(studio_id, params::Dict{String,<:Any})

List all the launch profiles a studio.

Arguments

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to be returned per request.

"nextToken": The token for the next set of results, or null if there are no more results.

"principalId": The principal ID.

"states": A list of states.

Main.Nimble.list_streaming_images — Method

list_streaming_images(studio_id)
list_streaming_images(studio_id, params::Dict{String,<:Any})

List the streaming image resources available to this studio. This list will contain both images provided by Amazon Web Services, as well as streaming images that you have created in your studio.

Arguments

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nextToken": The token for the next set of results, or null if there are no more results.

"owner": The owner.

Main.Nimble.list_streaming_sessions — Method

list_streaming_sessions(studio_id)
list_streaming_sessions(studio_id, params::Dict{String,<:Any})

Lists the streaming image resources in a studio.

Arguments

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"createdBy": The user ID of the user that created the streaming session.

"nextToken": The token for the next set of results, or null if there are no more results.

"ownedBy": The user ID of the user that owns the streaming session.

"sessionIds": A collection of session IDs.

Main.Nimble.list_studio_components — Method

list_studio_components(studio_id)
list_studio_components(studio_id, params::Dict{String,<:Any})

Lists the StudioComponents in a studio.

Arguments

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to be returned per request.

"nextToken": The token for the next set of results, or null if there are no more results.

"states": A list of states.

"types": The types.

Main.Nimble.list_studio_members — Method

list_studio_members(studio_id)
list_studio_members(studio_id, params::Dict{String,<:Any})

Get all users in a given studio membership.

Arguments

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to be returned per request.

"nextToken": The token for the next set of results, or null if there are no more results.

Main.Nimble.list_studios — Method

list_studios()
list_studios(params::Dict{String,<:Any})

List studios in your Amazon Web Services account in the requested Amazon Web Services Region.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nextToken": The token for the next set of results, or null if there are no more results.

Main.Nimble.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Gets the tags for a resource, given its Amazon Resource Names (ARN). This operation supports ARNs for all resource types in Nimble Studio that support tags, including studio, studio component, launch profile, streaming image, and streaming session. All resources that can be tagged will contain an ARN property, so you do not have to create this ARN yourself.

Arguments

resource_arn: The ARN of the target resource for tagging operations.

Main.Nimble.put_launch_profile_members — Method

put_launch_profile_members(identity_store_id, launch_profile_id, members, studio_id)
put_launch_profile_members(identity_store_id, launch_profile_id, members, studio_id, params::Dict{String,<:Any})

Add/update users with given persona to launch profile membership.

Arguments

identity_store_id: The ID of the identity store.

launch_profile_id: The launch profile ID.

members: A list of members.

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"X-Amz-Client-Token": To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

Main.Nimble.put_studio_members — Method

put_studio_members(identity_store_id, members, studio_id)
put_studio_members(identity_store_id, members, studio_id, params::Dict{String,<:Any})

Add/update users with given persona to studio membership.

Arguments

identity_store_id: The ID of the identity store.

members: A list of members.

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"X-Amz-Client-Token": To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

Main.Nimble.start_studio_ssoconfiguration_repair — Method

start_studio_ssoconfiguration_repair(studio_id)
start_studio_ssoconfiguration_repair(studio_id, params::Dict{String,<:Any})

Repairs the SSO configuration for a given studio. If the studio has a valid Amazon Web Services SSO configuration currently associated with it, this operation will fail with a validation error. If the studio does not have a valid Amazon Web Services SSO configuration currently associated with it, then a new Amazon Web Services SSO application is created for the studio and the studio is changed to the READY state. After the Amazon Web Services SSO application is repaired, you must use the Amazon Nimble Studio console to add administrators and users to your studio.

Arguments

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"X-Amz-Client-Token": To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

Main.Nimble.tag_resource — Method

tag_resource(resource_arn)
tag_resource(resource_arn, params::Dict{String,<:Any})

Creates tags for a resource, given its ARN.

Arguments

resource_arn: The ARN of the target resource for tagging operations.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"tags": A collection of labels, in the form of key:value pairs, that apply to this resource.

Main.Nimble.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Deletes the tags for a resource.

Arguments

resource_arn: The ARN of the target resource for tagging operations.

tag_keys: An array of tag keys to delete.

Main.Nimble.update_launch_profile — Method

update_launch_profile(launch_profile_id, studio_id)
update_launch_profile(launch_profile_id, studio_id, params::Dict{String,<:Any})

Update a launch profile.

Arguments

launch_profile_id: The launch profile ID.

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"X-Amz-Client-Token": To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

"description": The description.

"launchProfileProtocolVersions": The version number of the protocol that is used by the launch profile. The only valid version is "2021-03-31".

"name": The name for the launch profile.

"streamConfiguration": A configuration for a streaming session.

"studioComponentIds": Unique identifiers for a collection of studio components that can be used with this launch profile.

Main.Nimble.update_launch_profile_member — Method

update_launch_profile_member(launch_profile_id, persona, principal_id, studio_id)
update_launch_profile_member(launch_profile_id, persona, principal_id, studio_id, params::Dict{String,<:Any})

Update a user persona in launch profile membership.

Arguments

launch_profile_id: The launch profile ID.

persona: The persona.

principal_id: The principal ID.

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"X-Amz-Client-Token": To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

Main.Nimble.update_streaming_image — Method

update_streaming_image(streaming_image_id, studio_id)
update_streaming_image(streaming_image_id, studio_id, params::Dict{String,<:Any})

Update streaming image.

Arguments

streaming_image_id: The streaming image ID.

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"X-Amz-Client-Token": To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

"description": The description.

"name": The name for the streaming image.

Main.Nimble.update_studio — Method

update_studio(studio_id)
update_studio(studio_id, params::Dict{String,<:Any})

Update a Studio resource. Currently, this operation only supports updating the displayName of your studio.

Arguments

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"X-Amz-Client-Token": To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

"adminRoleArn": The IAM role that Studio Admins will assume when logging in to the Nimble Studio portal.

"displayName": A friendly name for the studio.

"userRoleArn": The IAM role that Studio Users will assume when logging in to the Nimble Studio portal.

Main.Nimble.update_studio_component — Method

update_studio_component(studio_component_id, studio_id)
update_studio_component(studio_component_id, studio_id, params::Dict{String,<:Any})

Updates a studio component resource.

Arguments

studio_component_id: The studio component ID.

studio_id: The studio ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"X-Amz-Client-Token": To make an idempotent API request using one of these actions, specify a client token in the request. You should not reuse the same client token for other API requests. If you retry a request that completed successfully using the same client token and the same parameters, the retry succeeds without performing any further actions. If you retry a successful request using the same client token, but one or more of the parameters are different, the retry fails with a ValidationException error.

"configuration": The configuration of the studio component, based on component type.

"description": The description.

"ec2SecurityGroupIds": The EC2 security groups that control access to the studio component.

"initializationScripts": Initialization scripts for studio components.

"name": The name for the studio component.

"scriptParameters": Parameters for the studio component scripts.

"subtype": The specific subtype of a studio component.

"type": The type of the studio component.