Notifications

This page documents function available when using the Notifications module, created with @service Notifications.

Index

Main.Notifications.associate_channel

Main.Notifications.associate_managed_notification_account_contact

Main.Notifications.associate_managed_notification_additional_channel

Main.Notifications.associate_organizational_unit

Main.Notifications.create_event_rule

Main.Notifications.create_notification_configuration

Main.Notifications.delete_event_rule

Main.Notifications.delete_notification_configuration

Main.Notifications.deregister_notification_hub

Main.Notifications.disable_notifications_access_for_organization

Main.Notifications.disassociate_channel

Main.Notifications.disassociate_managed_notification_account_contact

Main.Notifications.disassociate_managed_notification_additional_channel

Main.Notifications.disassociate_organizational_unit

Main.Notifications.enable_notifications_access_for_organization

Main.Notifications.get_event_rule

Main.Notifications.get_managed_notification_child_event

Main.Notifications.get_managed_notification_configuration

Main.Notifications.get_managed_notification_event

Main.Notifications.get_notification_configuration

Main.Notifications.get_notification_event

Main.Notifications.get_notifications_access_for_organization

Main.Notifications.list_channels

Main.Notifications.list_event_rules

Main.Notifications.list_managed_notification_channel_associations

Main.Notifications.list_managed_notification_child_events

Main.Notifications.list_managed_notification_configurations

Main.Notifications.list_managed_notification_events

Main.Notifications.list_member_accounts

Main.Notifications.list_notification_configurations

Main.Notifications.list_notification_events

Main.Notifications.list_notification_hubs

Main.Notifications.list_organizational_units

Main.Notifications.list_tags_for_resource

Main.Notifications.register_notification_hub

Main.Notifications.tag_resource

Main.Notifications.untag_resource

Main.Notifications.update_event_rule

Main.Notifications.update_notification_configuration

Documentation

Main.Notifications.associate_channel — Function

associate_channel(arn, notification_configuration_arn)
associate_channel(arn, notification_configuration_arn, params::Dict{String,<:Any})

Associates a delivery Channel with a particular NotificationConfiguration. Supported Channels include Amazon Q Developer in chat applications, the Console Mobile Application, and emails (notifications-contacts).

Arguments

arn: The Amazon Resource Name (ARN) of the Channel to associate with the NotificationConfiguration.

Supported ARNs include Amazon Q Developer in chat applications, the Console Mobile Application, and notifications-contacts.

notification_configuration_arn: The ARN of the NotificationConfiguration to associate with the Channel.

Main.Notifications.associate_managed_notification_account_contact — Function

associate_managed_notification_account_contact(contact_identifier, managed_notification_configuration_arn)
associate_managed_notification_account_contact(contact_identifier, managed_notification_configuration_arn, params::Dict{String,<:Any})

Associates an Account Contact with a particular ManagedNotificationConfiguration.

Arguments

contact_identifier: A unique value of an Account Contact Type to associate with the ManagedNotificationConfiguration.

managed_notification_configuration_arn: The Amazon Resource Name (ARN) of the ManagedNotificationConfiguration to associate with the Account Contact.

Main.Notifications.associate_managed_notification_additional_channel — Function

associate_managed_notification_additional_channel(channel_arn, managed_notification_configuration_arn)
associate_managed_notification_additional_channel(channel_arn, managed_notification_configuration_arn, params::Dict{String,<:Any})

Associates an additional Channel with a particular ManagedNotificationConfiguration.

Supported Channels include Amazon Q Developer in chat applications, the Console Mobile Application, and emails (notifications-contacts).

Arguments

channel_arn: The Amazon Resource Name (ARN) of the Channel to associate with the ManagedNotificationConfiguration.

Supported ARNs include Amazon Q Developer in chat applications, the Console Mobile Application, and email (notifications-contacts).

managed_notification_configuration_arn: The Amazon Resource Name (ARN) of the ManagedNotificationConfiguration to associate with the additional Channel.

Main.Notifications.associate_organizational_unit — Function

associate_organizational_unit(notification_configuration_arn, organizational_unit_id)
associate_organizational_unit(notification_configuration_arn, organizational_unit_id, params::Dict{String,<:Any})

Associates an organizational unit with a notification configuration.

Arguments

notification_configuration_arn: The Amazon Resource Name (ARN) of the notification configuration to associate with the organizational unit.

organizational_unit_id: The unique identifier of the organizational unit to associate.

Main.Notifications.create_event_rule — Function

create_event_rule(event_type, notification_configuration_arn, regions, source)
create_event_rule(event_type, notification_configuration_arn, regions, source, params::Dict{String,<:Any})

Creates an EventRule that is associated with a specified NotificationConfiguration.

Arguments

event_type: The event type to match.

Must match one of the valid Amazon EventBridge event types. For example, EC2 Instance State-change Notification and Amazon CloudWatch Alarm State Change. For more information, see Event delivery from Amazon Web Services services in the Amazon EventBridge User Guide.

notification_configuration_arn: The Amazon Resource Name (ARN) of the NotificationConfiguration associated with this EventRule.

regions: A list of Amazon Web Services Regions that send events to this EventRule.

source: The matched event source.

Must match one of the valid EventBridge sources. Only Amazon Web Services service sourced events are supported. For example, aws.ec2 and aws.cloudwatch. For more information, see Event delivery from Amazon Web Services services in the Amazon EventBridge User Guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"eventPattern": An additional event pattern used to further filter the events this EventRule receives.

For more information, see Amazon EventBridge event patterns in the Amazon EventBridge User Guide.

Main.Notifications.create_notification_configuration — Function

create_notification_configuration(description, name)
create_notification_configuration(description, name, params::Dict{String,<:Any})

Creates a new NotificationConfiguration.

Arguments

description: The description of the NotificationConfiguration.

name: The name of the NotificationConfiguration. Supports RFC 3986's unreserved characters.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"aggregationDuration": The aggregation preference of the NotificationConfiguration.

Values:

LONG

Aggregate notifications for long periods of time (12 hours).

SHORT

Aggregate notifications for short periods of time (5 minutes).

NONE

Don't aggregate notifications.

"tags": A map of tags assigned to a resource. A tag is a string-to-string map of key- value pairs.

Main.Notifications.delete_event_rule — Function

delete_event_rule(arn)
delete_event_rule(arn, params::Dict{String,<:Any})

Deletes an EventRule.

Arguments

arn: The Amazon Resource Name (ARN) of the EventRule to delete.

Main.Notifications.delete_notification_configuration — Function

delete_notification_configuration(arn)
delete_notification_configuration(arn, params::Dict{String,<:Any})

Deletes a NotificationConfiguration.

Arguments

arn: The Amazon Resource Name (ARN) of the NotificationConfiguration to delete.

Main.Notifications.deregister_notification_hub — Function

deregister_notification_hub(notification_hub_region)
deregister_notification_hub(notification_hub_region, params::Dict{String,<:Any})

Deregisters a NotificationConfiguration in the specified Region.

Note

You can't deregister the last NotificationHub in the account. NotificationEvents stored in the deregistered NotificationConfiguration are no longer be visible. Recreating a new NotificationConfiguration in the same Region restores access to those NotificationEvents.

Arguments

notification_hub_region: The NotificationConfiguration Region.

Main.Notifications.disable_notifications_access_for_organization — Function

disable_notifications_access_for_organization()
disable_notifications_access_for_organization(params::Dict{String,<:Any})

Disables service trust between User Notifications and Amazon Web Services Organizations.

Main.Notifications.disassociate_channel — Function

disassociate_channel(arn, notification_configuration_arn)
disassociate_channel(arn, notification_configuration_arn, params::Dict{String,<:Any})

Disassociates a Channel from a specified NotificationConfiguration. Supported Channels include Amazon Q Developer in chat applications, the Console Mobile Application, and emails (notifications-contacts).

Arguments

arn: The Amazon Resource Name (ARN) of the Channel to disassociate.

notification_configuration_arn: The ARN of the NotificationConfiguration to disassociate.

Main.Notifications.disassociate_managed_notification_account_contact — Function

disassociate_managed_notification_account_contact(contact_identifier, managed_notification_configuration_arn)
disassociate_managed_notification_account_contact(contact_identifier, managed_notification_configuration_arn, params::Dict{String,<:Any})

Disassociates an Account Contact with a particular ManagedNotificationConfiguration.

Arguments

contact_identifier: The unique value of an Account Contact Type to associate with the ManagedNotificationConfiguration.

managed_notification_configuration_arn: The Amazon Resource Name (ARN) of the ManagedNotificationConfiguration to associate with the Account Contact.

Main.Notifications.disassociate_managed_notification_additional_channel — Function

disassociate_managed_notification_additional_channel(channel_arn, managed_notification_configuration_arn)
disassociate_managed_notification_additional_channel(channel_arn, managed_notification_configuration_arn, params::Dict{String,<:Any})

Disassociates an additional Channel from a particular ManagedNotificationConfiguration.

Supported Channels include Amazon Q Developer in chat applications, the Console Mobile Application, and emails (notifications-contacts).

Arguments

channel_arn: The Amazon Resource Name (ARN) of the Channel to associate with the ManagedNotificationConfiguration.

managed_notification_configuration_arn: The Amazon Resource Name (ARN) of the Managed Notification Configuration to associate with the additional Channel.

Main.Notifications.disassociate_organizational_unit — Function

disassociate_organizational_unit(notification_configuration_arn, organizational_unit_id)
disassociate_organizational_unit(notification_configuration_arn, organizational_unit_id, params::Dict{String,<:Any})

Removes the association between an organizational unit and a notification configuration.

Arguments

notification_configuration_arn: The Amazon Resource Name (ARN) of the notification configuration to disassociate from the organizational unit.

organizational_unit_id: The unique identifier of the organizational unit to disassociate.

Main.Notifications.enable_notifications_access_for_organization — Function

enable_notifications_access_for_organization()
enable_notifications_access_for_organization(params::Dict{String,<:Any})

Enables service trust between User Notifications and Amazon Web Services Organizations.

Main.Notifications.get_event_rule — Function

get_event_rule(arn)
get_event_rule(arn, params::Dict{String,<:Any})

Returns a specified EventRule.

Arguments

arn: The Amazon Resource Name (ARN) of the EventRule to return.

Main.Notifications.get_managed_notification_child_event — Function

get_managed_notification_child_event(arn)
get_managed_notification_child_event(arn, params::Dict{String,<:Any})

Returns the child event of a specific given ManagedNotificationEvent.

Arguments

arn: The Amazon Resource Name (ARN) of the ManagedNotificationChildEvent to return.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"locale": The locale code of the language used for the retrieved ManagedNotificationChildEvent. The default locale is English en_US.

Main.Notifications.get_managed_notification_configuration — Function

get_managed_notification_configuration(arn)
get_managed_notification_configuration(arn, params::Dict{String,<:Any})

Returns a specified ManagedNotificationConfiguration.

Arguments

arn: The Amazon Resource Name (ARN) of the ManagedNotificationConfiguration to return.

Main.Notifications.get_managed_notification_event — Function

get_managed_notification_event(arn)
get_managed_notification_event(arn, params::Dict{String,<:Any})

Returns a specified ManagedNotificationEvent.

Arguments

arn: The Amazon Resource Name (ARN) of the ManagedNotificationEvent to return.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"locale": The locale code of the language used for the retrieved ManagedNotificationEvent. The default locale is English (en_US).

Main.Notifications.get_notification_configuration — Function

get_notification_configuration(arn)
get_notification_configuration(arn, params::Dict{String,<:Any})

Returns a specified NotificationConfiguration.

Arguments

arn: The Amazon Resource Name (ARN) of the NotificationConfiguration to return.

Main.Notifications.get_notification_event — Function

get_notification_event(arn)
get_notification_event(arn, params::Dict{String,<:Any})

Returns a specified NotificationEvent.

Important

User Notifications stores notifications in the individual Regions you register as notification hubs and the Region of the source event rule. GetNotificationEvent only returns notifications stored in the same Region in which the action is called. User Notifications doesn't backfill notifications to new Regions selected as notification hubs. For this reason, we recommend that you make calls in your oldest registered notification hub. For more information, see Notification hubs in the Amazon Web Services User Notifications User Guide.

Arguments

arn: The Amazon Resource Name (ARN) of the NotificationEvent to return.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"locale": The locale code of the language used for the retrieved NotificationEvent. The default locale is English en_US.

Main.Notifications.get_notifications_access_for_organization — Function

get_notifications_access_for_organization()
get_notifications_access_for_organization(params::Dict{String,<:Any})

Returns the AccessStatus of Service Trust Enablement for User Notifications and Amazon Web Services Organizations.

Main.Notifications.list_channels — Function

list_channels(notification_configuration_arn)
list_channels(notification_configuration_arn, params::Dict{String,<:Any})

Returns a list of Channels for a NotificationConfiguration.

Arguments

notification_configuration_arn: The Amazon Resource Name (ARN) of the NotificationConfiguration.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to be returned in this call. The default value is 20.

"nextToken": The start token for paginated calls. Retrieved from the response of a previous ListNotificationEvents call. NextToken uses Base64 encoding.

Main.Notifications.list_event_rules — Function

list_event_rules(notification_configuration_arn)
list_event_rules(notification_configuration_arn, params::Dict{String,<:Any})

Returns a list of EventRules according to specified filters, in reverse chronological order (newest first).

Arguments

notification_configuration_arn: The Amazon Resource Name (ARN) of the NotificationConfiguration.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to be returned in this call. The default value is 20.

"nextToken": The start token for paginated calls. Retrieved from the response of a previous ListEventRules call. Next token uses Base64 encoding.

Main.Notifications.list_managed_notification_channel_associations — Function

list_managed_notification_channel_associations(managed_notification_configuration_arn)
list_managed_notification_channel_associations(managed_notification_configuration_arn, params::Dict{String,<:Any})

Returns a list of Account contacts and Channels associated with a ManagedNotificationConfiguration, in paginated format.

Arguments

managed_notification_configuration_arn: The Amazon Resource Name (ARN) of the ManagedNotificationConfiguration to match.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to be returned in this call. Defaults to 20.

"nextToken": The start token for paginated calls. Retrieved from the response of a previous ListManagedNotificationChannelAssociations call.

Main.Notifications.list_managed_notification_child_events — Function

list_managed_notification_child_events(aggregate_managed_notification_event_arn)
list_managed_notification_child_events(aggregate_managed_notification_event_arn, params::Dict{String,<:Any})

Returns a list of ManagedNotificationChildEvents for a specified aggregate ManagedNotificationEvent, ordered by creation time in reverse chronological order (newest first).

Arguments

aggregate_managed_notification_event_arn: The Amazon Resource Name (ARN) of the ManagedNotificationEvent.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"endTime": Latest time of events to return from this call.

"locale": The locale code of the language used for the retrieved NotificationEvent. The default locale is English.en_US.

"maxResults": The maximum number of results to be returned in this call. Defaults to 20.

"nextToken": The start token for paginated calls. Retrieved from the response of a previous ListManagedNotificationChannelAssociations call. Next token uses Base64 encoding.

"organizationalUnitId": The identifier of the Amazon Web Services Organizations organizational unit (OU) associated with the Managed Notification Child Events.

"relatedAccount": The Amazon Web Services account ID associated with the Managed Notification Child Events.

"startTime": The earliest time of events to return from this call.

Main.Notifications.list_managed_notification_configurations — Function

list_managed_notification_configurations()
list_managed_notification_configurations(params::Dict{String,<:Any})

Returns a list of Managed Notification Configurations according to specified filters, ordered by creation time in reverse chronological order (newest first).

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"channelIdentifier": The identifier or ARN of the notification channel to filter configurations by.

"maxResults": The maximum number of results to be returned in this call. Defaults to 20.

"nextToken": The start token for paginated calls. Retrieved from the response of a previous ListManagedNotificationChannelAssociations call. Next token uses Base64 encoding.

Main.Notifications.list_managed_notification_events — Function

list_managed_notification_events()
list_managed_notification_events(params::Dict{String,<:Any})

Returns a list of Managed Notification Events according to specified filters, ordered by creation time in reverse chronological order (newest first).

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"endTime": Latest time of events to return from this call.

"locale": The locale code of the language used for the retrieved NotificationEvent. The default locale is English (en_US).

"maxResults": The maximum number of results to be returned in this call. Defaults to 20.

"nextToken": The start token for paginated calls. Retrieved from the response of a previous ListManagedNotificationChannelAssociations call. Next token uses Base64 encoding.

"organizationalUnitId": The Organizational Unit Id that an Amazon Web Services account belongs to.

"relatedAccount": The Amazon Web Services account ID associated with the Managed Notification Events.

"source": The Amazon Web Services service the event originates from. For example aws.cloudwatch.

"startTime": The earliest time of events to return from this call.

Main.Notifications.list_member_accounts — Function

list_member_accounts(notification_configuration_arn)
list_member_accounts(notification_configuration_arn, params::Dict{String,<:Any})

Returns a list of member accounts associated with a notification configuration.

Arguments

notification_configuration_arn: The Amazon Resource Name (ARN) of the notification configuration used to filter the member accounts.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return in a single call. Valid values are 1-100.

"memberAccount": The member account identifier used to filter the results.

"nextToken": The token for the next page of results. Use the value returned in the previous response.

"organizationalUnitId": The organizational unit ID used to filter the member accounts.

"status": The status used to filter the member accounts.

Main.Notifications.list_notification_configurations — Function

list_notification_configurations()
list_notification_configurations(params::Dict{String,<:Any})

Returns a list of abbreviated NotificationConfigurations according to specified filters, in reverse chronological order (newest first).

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"channelArn": The Amazon Resource Name (ARN) of the Channel to match.

"eventRuleSource": The matched event source.

Must match one of the valid EventBridge sources. Only Amazon Web Services service sourced events are supported. For example, aws.ec2 and aws.cloudwatch. For more information, see Event delivery from Amazon Web Services services in the Amazon EventBridge User Guide.

"maxResults": The maximum number of results to be returned in this call. Defaults to 20.

"nextToken": The start token for paginated calls. Retrieved from the response of a previous ListEventRules call. Next token uses Base64 encoding.

"status": The NotificationConfiguration status to match.

Values:

ACTIVE

All EventRules are ACTIVE and any call can be run.

PARTIALLY_ACTIVE

Some EventRules are ACTIVE and some are INACTIVE. Any call can be run.

Any call can be run.

INACTIVE

All EventRules are INACTIVE and any call can be run.

DELETING

This NotificationConfiguration is being deleted.

Only GET and LIST calls can be run.

"subtype": The subtype used to filter the notification configurations in the request.

Main.Notifications.list_notification_events — Function

list_notification_events()
list_notification_events(params::Dict{String,<:Any})

Returns a list of NotificationEvents according to specified filters, in reverse chronological order (newest first).

Important

User Notifications stores notifications in the individual Regions you register as notification hubs and the Region of the source event rule. ListNotificationEvents only returns notifications stored in the same Region in which the action is called. User Notifications doesn't backfill notifications to new Regions selected as notification hubs. For this reason, we recommend that you make calls in your oldest registered notification hub. For more information, see Notification hubs in the Amazon Web Services User Notifications User Guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"aggregateNotificationEventArn": The Amazon Resource Name (ARN) of the aggregatedNotificationEventArn to match.

"endTime": Latest time of events to return from this call.

"includeChildEvents": Include aggregated child events in the result.

"locale": The locale code of the language used for the retrieved NotificationEvent. The default locale is English (en_US).

"maxResults": The maximum number of results to be returned in this call. Defaults to 20.

"nextToken": The start token for paginated calls. Retrieved from the response of a previous ListEventRules call. Next token uses Base64 encoding.

"organizationalUnitId": The unique identifier of the organizational unit used to filter notification events.

"source": The matched event source.

Must match one of the valid EventBridge sources. Only Amazon Web Services service sourced events are supported. For example, aws.ec2 and aws.cloudwatch. For more information, see Event delivery from Amazon Web Services services in the Amazon EventBridge User Guide.

"startTime": The earliest time of events to return from this call.

Main.Notifications.list_notification_hubs — Function

list_notification_hubs()
list_notification_hubs(params::Dict{String,<:Any})

Returns a list of NotificationHubs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of records to list in a single response.

"nextToken": A pagination token. Set to null to start listing notification hubs from the start.

Main.Notifications.list_organizational_units — Function

list_organizational_units(notification_configuration_arn)
list_organizational_units(notification_configuration_arn, params::Dict{String,<:Any})

Returns a list of organizational units associated with a notification configuration.

Arguments

notification_configuration_arn: The Amazon Resource Name (ARN) of the notification configuration used to filter the organizational units.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of organizational units to return in a single call. Valid values are 1-100.

"nextToken": The token for the next page of results. Use the value returned in the previous response.

Main.Notifications.list_tags_for_resource — Function

list_tags_for_resource(arn)
list_tags_for_resource(arn, params::Dict{String,<:Any})

Returns a list of tags for a specified Amazon Resource Name (ARN).

For more information, see Tagging your Amazon Web Services resources in the Tagging Amazon Web Services Resources User Guide.

Note

This is only supported for NotificationConfigurations.

Arguments

arn: The Amazon Resource Name (ARN) to use to list tags.

Main.Notifications.register_notification_hub — Function

register_notification_hub(notification_hub_region)
register_notification_hub(notification_hub_region, params::Dict{String,<:Any})

Registers a NotificationConfiguration in the specified Region.

There is a maximum of one NotificationConfiguration per Region. You can have a maximum of 3 NotificationHub resources at a time.

Arguments

notification_hub_region: The Region of the NotificationHub.

Main.Notifications.tag_resource — Function

tag_resource(arn, tags)
tag_resource(arn, tags, params::Dict{String,<:Any})

Tags the resource with a tag key and value.

For more information, see Tagging your Amazon Web Services resources in the Tagging Amazon Web Services Resources User Guide.

Note

This is only supported for NotificationConfigurations.

Arguments

arn: The Amazon Resource Name (ARN) to use to tag a resource.

tags: A map of tags assigned to a resource. A tag is a string-to-string map of key-value pairs.

Main.Notifications.untag_resource — Function

untag_resource(arn, tag_keys)
untag_resource(arn, tag_keys, params::Dict{String,<:Any})

Untags a resource with a specified Amazon Resource Name (ARN).

For more information, see Tagging your Amazon Web Services resources in the Tagging Amazon Web Services Resources User Guide.

Arguments

arn: The Amazon Resource Name (ARN) to use to untag a resource.

tag_keys: The tag keys to use to untag a resource.

Main.Notifications.update_event_rule — Function

update_event_rule(arn)
update_event_rule(arn, params::Dict{String,<:Any})

Updates an existing EventRule.

Arguments

arn: The Amazon Resource Name (ARN) to use to update the EventRule.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"eventPattern": An additional event pattern used to further filter the events this EventRule receives.

For more information, see Amazon EventBridge event patterns in the Amazon EventBridge User Guide.

"regions": A list of Amazon Web Services Regions that sends events to this EventRule.

Main.Notifications.update_notification_configuration — Function

update_notification_configuration(arn)
update_notification_configuration(arn, params::Dict{String,<:Any})

Updates a NotificationConfiguration.

Arguments

arn: The Amazon Resource Name (ARN) used to update the NotificationConfiguration.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"aggregationDuration": The aggregation preference of the NotificationConfiguration.

Values:

LONG

Aggregate notifications for long periods of time (12 hours).

SHORT

Aggregate notifications for short periods of time (5 minutes).

NONE

Don't aggregate notifications.

"description": The description of the NotificationConfiguration.

"name": The name of the NotificationConfiguration.