Odb

This page documents function available when using the Odb module, created with @service Odb.

Index

Main.Odb.accept_marketplace_registration

Main.Odb.associate_iam_role_to_resource

Main.Odb.create_cloud_autonomous_vm_cluster

Main.Odb.create_cloud_exadata_infrastructure

Main.Odb.create_cloud_vm_cluster

Main.Odb.create_odb_network

Main.Odb.create_odb_peering_connection

Main.Odb.delete_cloud_autonomous_vm_cluster

Main.Odb.delete_cloud_exadata_infrastructure

Main.Odb.delete_cloud_vm_cluster

Main.Odb.delete_odb_network

Main.Odb.delete_odb_peering_connection

Main.Odb.disassociate_iam_role_from_resource

Main.Odb.get_cloud_autonomous_vm_cluster

Main.Odb.get_cloud_exadata_infrastructure

Main.Odb.get_cloud_exadata_infrastructure_unallocated_resources

Main.Odb.get_cloud_vm_cluster

Main.Odb.get_db_node

Main.Odb.get_db_server

Main.Odb.get_oci_onboarding_status

Main.Odb.get_odb_network

Main.Odb.get_odb_peering_connection

Main.Odb.initialize_service

Main.Odb.list_autonomous_virtual_machines

Main.Odb.list_cloud_autonomous_vm_clusters

Main.Odb.list_cloud_exadata_infrastructures

Main.Odb.list_cloud_vm_clusters

Main.Odb.list_db_nodes

Main.Odb.list_db_servers

Main.Odb.list_db_system_shapes

Main.Odb.list_gi_versions

Main.Odb.list_odb_networks

Main.Odb.list_odb_peering_connections

Main.Odb.list_system_versions

Main.Odb.list_tags_for_resource

Main.Odb.reboot_db_node

Main.Odb.start_db_node

Main.Odb.stop_db_node

Main.Odb.tag_resource

Main.Odb.untag_resource

Main.Odb.update_cloud_exadata_infrastructure

Main.Odb.update_odb_network

Main.Odb.update_odb_peering_connection

Documentation

Main.Odb.accept_marketplace_registration — Function

accept_marketplace_registration(marketplace_registration_token)
accept_marketplace_registration(marketplace_registration_token, params::Dict{String,<:Any})

Registers the Amazon Web Services Marketplace token for your Amazon Web Services account to activate your Oracle Database@Amazon Web Services subscription.

Arguments

marketplace_registration_token: The registration token that's generated by Amazon Web Services Marketplace and sent to Oracle Database@Amazon Web Services.

Main.Odb.associate_iam_role_to_resource — Function

associate_iam_role_to_resource(aws_integration, iam_role_arn, resource_arn)
associate_iam_role_to_resource(aws_integration, iam_role_arn, resource_arn, params::Dict{String,<:Any})

Associates an Amazon Web Services Identity and Access Management (IAM) service role with a specified resource to enable Amazon Web Services service integration.

Arguments

aws_integration: The Amazon Web Services integration configuration settings for the Amazon Web Services Identity and Access Management (IAM) service role association.

iam_role_arn: The Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) service role to associate with the resource.

resource_arn: The Amazon Resource Name (ARN) of the target resource to associate with the Amazon Web Services Identity and Access Management (IAM) service role.

Main.Odb.create_cloud_autonomous_vm_cluster — Function

create_cloud_autonomous_vm_cluster(autonomous_data_storage_size_in_tbs, cloud_exadata_infrastructure_id, cpu_core_count_per_node, display_name, memory_per_oracle_compute_unit_in_gbs, odb_network_id, total_container_databases)
create_cloud_autonomous_vm_cluster(autonomous_data_storage_size_in_tbs, cloud_exadata_infrastructure_id, cpu_core_count_per_node, display_name, memory_per_oracle_compute_unit_in_gbs, odb_network_id, total_container_databases, params::Dict{String,<:Any})

Creates a new Autonomous VM cluster in the specified Exadata infrastructure.

Arguments

autonomous_data_storage_size_in_tbs: The data disk group size to be allocated for Autonomous Databases, in terabytes (TB).

cloud_exadata_infrastructure_id: The unique identifier of the Exadata infrastructure where the VM cluster will be created.

cpu_core_count_per_node: The number of CPU cores to be enabled per VM cluster node.

display_name: The display name for the Autonomous VM cluster. The name does not need to be unique.

memory_per_oracle_compute_unit_in_gbs: The amount of memory to be allocated per OCPU, in GB.

odb_network_id: The unique identifier of the ODB network to be used for the VM cluster.

total_container_databases: The total number of Autonomous CDBs that you can create in the Autonomous VM cluster.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A client-provided token to ensure idempotency of the request.

"dbServers": The list of database servers to be used for the Autonomous VM cluster.

"description": A user-provided description of the Autonomous VM cluster.

"isMtlsEnabledVmCluster": Specifies whether to enable mutual TLS (mTLS) authentication for the Autonomous VM cluster.

"licenseModel": The Oracle license model to apply to the Autonomous VM cluster.

"maintenanceWindow": The scheduling details for the maintenance window. Patching and system updates take place during the maintenance window.

"scanListenerPortNonTls": The SCAN listener port for non-TLS (TCP) protocol.

"scanListenerPortTls": The SCAN listener port for TLS (TCP) protocol.

"tags": Free-form tags for this resource. Each tag is a key-value pair with no predefined name, type, or namespace.

"timeZone": The time zone to use for the Autonomous VM cluster.

Main.Odb.create_cloud_exadata_infrastructure — Function

create_cloud_exadata_infrastructure(compute_count, display_name, shape, storage_count)
create_cloud_exadata_infrastructure(compute_count, display_name, shape, storage_count, params::Dict{String,<:Any})

Creates an Exadata infrastructure.

Arguments

compute_count: The number of database servers for the Exadata infrastructure. Valid values for this parameter depend on the shape. To get information about the minimum and maximum values, use the ListDbSystemShapes operation.

display_name: A user-friendly name for the Exadata infrastructure.

shape: The model name of the Exadata infrastructure. For the list of valid model names, use the ListDbSystemShapes operation.

storage_count: The number of storage servers to activate for this Exadata infrastructure. Valid values for this parameter depend on the shape. To get information about the minimum and maximum values, use the ListDbSystemShapes operation.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"availabilityZone": The name of the Availability Zone (AZ) where the Exadata infrastructure is located.

This operation requires that you specify a value for either availabilityZone or availabilityZoneId.

Example: us-east-1a

"availabilityZoneId": The AZ ID of the AZ where the Exadata infrastructure is located.

This operation requires that you specify a value for either availabilityZone or availabilityZoneId.

Example: use1-az1

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If you don't specify a client token, the Amazon Web Services SDK automatically generates a client token and uses it for the request to ensure idempotency. The client token is valid for up to 24 hours after it's first used.

"customerContactsToSendToOCI": The email addresses of contacts to receive notification from Oracle about maintenance updates for the Exadata infrastructure.

"databaseServerType": The database server model type of the Exadata infrastructure. For the list of valid model names, use the ListDbSystemShapes operation.

"maintenanceWindow": The maintenance window configuration for the Exadata Cloud infrastructure.

This allows you to define when maintenance operations such as patching and updates can be performed on the infrastructure.

"storageServerType": The storage server model type of the Exadata infrastructure. For the list of valid model names, use the ListDbSystemShapes operation.

"tags": The list of resource tags to apply to the Exadata infrastructure.

Main.Odb.create_cloud_vm_cluster — Function

create_cloud_vm_cluster(cloud_exadata_infrastructure_id, cpu_core_count, display_name, gi_version, hostname, odb_network_id, ssh_public_keys)
create_cloud_vm_cluster(cloud_exadata_infrastructure_id, cpu_core_count, display_name, gi_version, hostname, odb_network_id, ssh_public_keys, params::Dict{String,<:Any})

Creates a VM cluster on the specified Exadata infrastructure.

Arguments

cloud_exadata_infrastructure_id: The unique identifier of the Exadata infrastructure for this VM cluster.

cpu_core_count: The number of CPU cores to enable on the VM cluster.

display_name: A user-friendly name for the VM cluster.

gi_version: A valid software version of Oracle Grid Infrastructure (GI). To get the list of valid values, use the ListGiVersions operation and specify the shape of the Exadata infrastructure.

Example: 19.0.0.0

hostname: The host name for the VM cluster.

Constraints:

Can't be "localhost" or "hostname".

Can't contain "-version".

The maximum length of the combined hostname and domain is 63 characters.

The hostname must be unique within the subnet.

odb_network_id: The unique identifier of the ODB network for the VM cluster.

ssh_public_keys: The public key portion of one or more key pairs used for SSH access to the VM cluster.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If you don't specify a client token, the Amazon Web Services SDK automatically generates a client token and uses it for the request to ensure idempotency. The client token is valid for up to 24 hours after it's first used.

"clusterName": A name for the Grid Infrastructure cluster. The name isn't case sensitive.

"dataCollectionOptions": The set of preferences for the various diagnostic collection options for the VM cluster.

"dataStorageSizeInTBs": The size of the data disk group, in terabytes (TBs), to allocate for the VM cluster.

"dbNodeStorageSizeInGBs": The amount of local node storage, in gigabytes (GBs), to allocate for the VM cluster.

"dbServers": The list of database servers for the VM cluster.

"isLocalBackupEnabled": Specifies whether to enable database backups to local Exadata storage for the VM cluster.

"isSparseDiskgroupEnabled": Specifies whether to create a sparse disk group for the VM cluster.

"licenseModel": The Oracle license model to apply to the VM cluster.

Default: LICENSE_INCLUDED

"memorySizeInGBs": The amount of memory, in gigabytes (GBs), to allocate for the VM cluster.

"scanListenerPortTcp": The port number for TCP connections to the single client access name (SCAN) listener.

Valid values: 1024–8999 with the following exceptions: 2484, 6100, 6200, 7060, 7070, 7085, and 7879

Default: 1521

"systemVersion": The version of the operating system of the image for the VM cluster.

"tags": The list of resource tags to apply to the VM cluster.

"timeZone": The time zone for the VM cluster. For a list of valid values for time zone, you can check the options in the console.

Default: UTC

Main.Odb.create_odb_network — Function

create_odb_network(client_subnet_cidr, display_name)
create_odb_network(client_subnet_cidr, display_name, params::Dict{String,<:Any})

Creates an ODB network.

Arguments

client_subnet_cidr: The CIDR range of the client subnet for the ODB network.

Constraints:

Must not overlap with the CIDR range of the backup subnet.

Must not overlap with the CIDR ranges of the VPCs that are connected to the ODB network.

Must not use the following CIDR ranges that are reserved by OCI:

100.106.0.0/16 and 100.107.0.0/16

169.254.0.0/16

224.0.0.0 - 239.255.255.255

240.0.0.0 - 255.255.255.255

display_name: A user-friendly name for the ODB network.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"availabilityZone": The Amazon Web Services Availability Zone (AZ) where the ODB network is located.

This operation requires that you specify a value for either availabilityZone or availabilityZoneId.

"availabilityZoneId": The AZ ID of the AZ where the ODB network is located.

This operation requires that you specify a value for either availabilityZone or availabilityZoneId.

"backupSubnetCidr": The CIDR range of the backup subnet for the ODB network.

Constraints:

Must not overlap with the CIDR range of the client subnet.

Must not overlap with the CIDR ranges of the VPCs that are connected to the ODB network.

Must not use the following CIDR ranges that are reserved by OCI:

100.106.0.0/16 and 100.107.0.0/16

169.254.0.0/16

224.0.0.0 - 239.255.255.255

240.0.0.0 - 255.255.255.255

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If you don't specify a client token, the Amazon Web Services SDK automatically generates a client token and uses it for the request to ensure idempotency. The client token is valid for up to 24 hours after it's first used.

"crossRegionS3RestoreSourcesToEnable": The cross-Region Amazon S3 restore sources to enable for the ODB network.

"customDomainName": The domain name to use for the resources in the ODB network.

"defaultDnsPrefix": The DNS prefix to the default DNS domain name. The default DNS domain name is oraclevcn.com.

"kmsAccess": The Amazon Web Services Key Management Service (KMS) access configuration for the ODB network.

"kmsPolicyDocument": The Amazon Web Services Key Management Service (KMS) policy document that defines permissions for key usage within the ODB network.

"s3Access": Specifies the configuration for Amazon S3 access from the ODB network.

"s3PolicyDocument": Specifies the endpoint policy for Amazon S3 access from the ODB network.

"stsAccess": The Amazon Web Services Security Token Service (STS) access configuration for the ODB network.

"stsPolicyDocument": The Amazon Web Services Security Token Service (STS) policy document that defines permissions for token service usage within the ODB network.

"tags": The list of resource tags to apply to the ODB network.

"zeroEtlAccess": Specifies the configuration for Zero-ETL access from the ODB network.

Main.Odb.create_odb_peering_connection — Function

create_odb_peering_connection(odb_network_id, peer_network_id)
create_odb_peering_connection(odb_network_id, peer_network_id, params::Dict{String,<:Any})

Creates a peering connection between an ODB network and a VPC.

A peering connection enables private connectivity between the networks for application-tier communication.

Arguments

odb_network_id: The unique identifier of the ODB network that initiates the peering connection.

peer_network_id: The unique identifier of the peer network. This can be either a VPC ID or another ODB network ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": The client token for the ODB peering connection request.

Constraints:

Must be unique for each request.

"displayName": The display name for the ODB peering connection.

"peerNetworkCidrsToBeAdded": A list of CIDR blocks to add to the peering connection. These CIDR blocks define the IP address ranges that can communicate through the peering connection.

"peerNetworkRouteTableIds": The unique identifier of the VPC route table for which a route to the ODB network is automatically created during peering connection establishment.

"tags": The tags to assign to the ODB peering connection.

Main.Odb.delete_cloud_autonomous_vm_cluster — Function

delete_cloud_autonomous_vm_cluster(cloud_autonomous_vm_cluster_id)
delete_cloud_autonomous_vm_cluster(cloud_autonomous_vm_cluster_id, params::Dict{String,<:Any})

Deletes an Autonomous VM cluster.

Arguments

cloud_autonomous_vm_cluster_id: The unique identifier of the Autonomous VM cluster to delete.

Main.Odb.delete_cloud_exadata_infrastructure — Function

delete_cloud_exadata_infrastructure(cloud_exadata_infrastructure_id)
delete_cloud_exadata_infrastructure(cloud_exadata_infrastructure_id, params::Dict{String,<:Any})

Deletes the specified Exadata infrastructure. Before you use this operation, make sure to delete all of the VM clusters that are hosted on this Exadata infrastructure.

Arguments

cloud_exadata_infrastructure_id: The unique identifier of the Exadata infrastructure to delete.

Main.Odb.delete_cloud_vm_cluster — Function

delete_cloud_vm_cluster(cloud_vm_cluster_id)
delete_cloud_vm_cluster(cloud_vm_cluster_id, params::Dict{String,<:Any})

Deletes the specified VM cluster.

Arguments

cloud_vm_cluster_id: The unique identifier of the VM cluster to delete.

Main.Odb.delete_odb_network — Function

delete_odb_network(delete_associated_resources, odb_network_id)
delete_odb_network(delete_associated_resources, odb_network_id, params::Dict{String,<:Any})

Deletes the specified ODB network.

Arguments

delete_associated_resources: Specifies whether to delete associated OCI networking resources along with the ODB network.

odb_network_id: The unique identifier of the ODB network to delete.

Main.Odb.delete_odb_peering_connection — Function

delete_odb_peering_connection(odb_peering_connection_id)
delete_odb_peering_connection(odb_peering_connection_id, params::Dict{String,<:Any})

Deletes an ODB peering connection.

When you delete an ODB peering connection, the underlying VPC peering connection is also deleted.

Arguments

odb_peering_connection_id: The unique identifier of the ODB peering connection to delete.

Main.Odb.disassociate_iam_role_from_resource — Function

disassociate_iam_role_from_resource(aws_integration, iam_role_arn, resource_arn)
disassociate_iam_role_from_resource(aws_integration, iam_role_arn, resource_arn, params::Dict{String,<:Any})

Disassociates an Amazon Web Services Identity and Access Management (IAM) service role from a specified resource to disable Amazon Web Services service integration.

Arguments

aws_integration: The Amazon Web Services integration configuration settings for the Amazon Web Services Identity and Access Management (IAM) service role disassociation.

iam_role_arn: The Amazon Resource Name (ARN) of the Amazon Web Services Identity and Access Management (IAM) service role to disassociate from the resource.

resource_arn: The Amazon Resource Name (ARN) of the target resource to disassociate from the Amazon Web Services Identity and Access Management (IAM) service role.

Main.Odb.get_cloud_autonomous_vm_cluster — Function

get_cloud_autonomous_vm_cluster(cloud_autonomous_vm_cluster_id)
get_cloud_autonomous_vm_cluster(cloud_autonomous_vm_cluster_id, params::Dict{String,<:Any})

Gets information about a specific Autonomous VM cluster.

Arguments

cloud_autonomous_vm_cluster_id: The unique identifier of the Autonomous VM cluster to retrieve information about.

Main.Odb.get_cloud_exadata_infrastructure — Function

get_cloud_exadata_infrastructure(cloud_exadata_infrastructure_id)
get_cloud_exadata_infrastructure(cloud_exadata_infrastructure_id, params::Dict{String,<:Any})

Returns information about the specified Exadata infrastructure.

Arguments

cloud_exadata_infrastructure_id: The unique identifier of the Exadata infrastructure.

Main.Odb.get_cloud_exadata_infrastructure_unallocated_resources — Function

get_cloud_exadata_infrastructure_unallocated_resources(cloud_exadata_infrastructure_id)
get_cloud_exadata_infrastructure_unallocated_resources(cloud_exadata_infrastructure_id, params::Dict{String,<:Any})

Retrieves information about unallocated resources in a specified Cloud Exadata Infrastructure.

Arguments

cloud_exadata_infrastructure_id: The unique identifier of the Cloud Exadata infrastructure for which to retrieve unallocated resources.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"dbServers": The database servers to include in the unallocated resources query.

Main.Odb.get_cloud_vm_cluster — Function

get_cloud_vm_cluster(cloud_vm_cluster_id)
get_cloud_vm_cluster(cloud_vm_cluster_id, params::Dict{String,<:Any})

Returns information about the specified VM cluster.

Arguments

cloud_vm_cluster_id: The unique identifier of the VM cluster.

Main.Odb.get_db_node — Function

get_db_node(cloud_vm_cluster_id, db_node_id)
get_db_node(cloud_vm_cluster_id, db_node_id, params::Dict{String,<:Any})

Returns information about the specified DB node.

Arguments

cloud_vm_cluster_id: The unique identifier of the VM cluster that contains the DB node.

db_node_id: The unique identifier of the DB node to retrieve information about.

Main.Odb.get_db_server — Function

get_db_server(cloud_exadata_infrastructure_id, db_server_id)
get_db_server(cloud_exadata_infrastructure_id, db_server_id, params::Dict{String,<:Any})

Returns information about the specified database server.

Arguments

cloud_exadata_infrastructure_id: The unique identifier of the Oracle Exadata infrastructure that contains the database server.

db_server_id: The unique identifier of the database server to retrieve information about.

Main.Odb.get_oci_onboarding_status — Function

get_oci_onboarding_status()
get_oci_onboarding_status(params::Dict{String,<:Any})

Returns the tenancy activation link and onboarding status for your Amazon Web Services account.

Main.Odb.get_odb_network — Function

get_odb_network(odb_network_id)
get_odb_network(odb_network_id, params::Dict{String,<:Any})

Returns information about the specified ODB network.

Arguments

odb_network_id: The unique identifier of the ODB network.

Main.Odb.get_odb_peering_connection — Function

get_odb_peering_connection(odb_peering_connection_id)
get_odb_peering_connection(odb_peering_connection_id, params::Dict{String,<:Any})

Retrieves information about an ODB peering connection.

Arguments

odb_peering_connection_id: The unique identifier of the ODB peering connection to retrieve information about.

Main.Odb.initialize_service — Function

initialize_service()
initialize_service(params::Dict{String,<:Any})

Initializes the ODB service for the first time in an account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ociIdentityDomain": The Oracle Cloud Infrastructure (OCI) identity domain configuration for service initialization.

Main.Odb.list_autonomous_virtual_machines — Function

list_autonomous_virtual_machines(cloud_autonomous_vm_cluster_id)
list_autonomous_virtual_machines(cloud_autonomous_vm_cluster_id, params::Dict{String,<:Any})

Lists all Autonomous VMs in an Autonomous VM cluster.

Arguments

cloud_autonomous_vm_cluster_id: The unique identifier of the Autonomous VM cluster whose virtual machines you're listing.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of items to return per page.

"nextToken": The pagination token to continue listing from.

Main.Odb.list_cloud_autonomous_vm_clusters — Function

list_cloud_autonomous_vm_clusters()
list_cloud_autonomous_vm_clusters(params::Dict{String,<:Any})

Lists all Autonomous VM clusters in a specified Cloud Exadata infrastructure.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"cloudExadataInfrastructureId": The unique identifier of the Cloud Exadata Infrastructure that hosts the Autonomous VM clusters to be listed.

"maxResults": The maximum number of items to return per page.

"nextToken": The pagination token to continue listing from.

Main.Odb.list_cloud_exadata_infrastructures — Function

list_cloud_exadata_infrastructures()
list_cloud_exadata_infrastructures(params::Dict{String,<:Any})

Returns information about the Exadata infrastructures owned by your Amazon Web Services account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of items to return for this request. To get the next page of items, make another request with the token returned in the output.

Default: 10

"nextToken": The token returned from a previous paginated request. Pagination continues from the end of the items returned by the previous request.

Main.Odb.list_cloud_vm_clusters — Function

list_cloud_vm_clusters()
list_cloud_vm_clusters(params::Dict{String,<:Any})

Returns information about the VM clusters owned by your Amazon Web Services account or only the ones on the specified Exadata infrastructure.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"cloudExadataInfrastructureId": The unique identifier of the Oracle Exadata infrastructure.

"maxResults": The maximum number of items to return for this request. To get the next page of items, make another request with the token returned in the output.

Default: 10

"nextToken": The token returned from a previous paginated request. Pagination continues from the end of the items returned by the previous request.

Main.Odb.list_db_nodes — Function

list_db_nodes(cloud_vm_cluster_id)
list_db_nodes(cloud_vm_cluster_id, params::Dict{String,<:Any})

Returns information about the DB nodes for the specified VM cluster.

Arguments

cloud_vm_cluster_id: The unique identifier of the VM cluster.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of items to return for this request. To get the next page of items, make another request with the token returned in the output.

Default: 10

"nextToken": The token returned from a previous paginated request. Pagination continues from the end of the items returned by the previous request.

Main.Odb.list_db_servers — Function

list_db_servers(cloud_exadata_infrastructure_id)
list_db_servers(cloud_exadata_infrastructure_id, params::Dict{String,<:Any})

Returns information about the database servers that belong to the specified Exadata infrastructure.

Arguments

cloud_exadata_infrastructure_id: The unique identifier of the Oracle Exadata infrastructure.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of items to return for this request. To get the next page of items, make another request with the token returned in the output.

Default: 10

"nextToken": The token returned from a previous paginated request. Pagination continues from the end of the items returned by the previous request.

Main.Odb.list_db_system_shapes — Function

list_db_system_shapes()
list_db_system_shapes(params::Dict{String,<:Any})

Returns information about the shapes that are available for an Exadata infrastructure.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"availabilityZone": The logical name of the AZ, for example, us-east-1a. This name varies depending on the account.

"availabilityZoneId": The physical ID of the AZ, for example, use1-az4. This ID persists across accounts.

"maxResults": The maximum number of items to return for this request. To get the next page of items, make another request with the token returned in the output.

Default: 10

"nextToken": The token returned from a previous paginated request. Pagination continues from the end of the items returned by the previous request.

Main.Odb.list_gi_versions — Function

list_gi_versions()
list_gi_versions(params::Dict{String,<:Any})

Returns information about Oracle Grid Infrastructure (GI) software versions that are available for a VM cluster for the specified shape.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of items to return for this request. To get the next page of items, make another request with the token returned in the output.

Default: 10

"nextToken": The token returned from a previous paginated request. Pagination continues from the end of the items returned by the previous request.

"shape": The shape to return GI versions for. For a list of valid shapes, use the ListDbSystemShapes operation..

Main.Odb.list_odb_networks — Function

list_odb_networks()
list_odb_networks(params::Dict{String,<:Any})

Returns information about the ODB networks owned by your Amazon Web Services account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of items to return for this request. To get the next page of items, make another request with the token returned in the output.

Default: 10

"nextToken": The token returned from a previous paginated request. Pagination continues from the end of the items returned by the previous request.

Main.Odb.list_odb_peering_connections — Function

list_odb_peering_connections()
list_odb_peering_connections(params::Dict{String,<:Any})

Lists all ODB peering connections or those associated with a specific ODB network.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of ODB peering connections to return in the response.

Default: 20

Constraints:

Must be between 1 and 100.

"nextToken": The pagination token for the next page of ODB peering connections.

"odbNetworkId": The identifier of the ODB network to list peering connections for.

If not specified, lists all ODB peering connections in the account.

Main.Odb.list_system_versions — Function

list_system_versions(gi_version, shape)
list_system_versions(gi_version, shape, params::Dict{String,<:Any})

Returns information about the system versions that are available for a VM cluster for the specified giVersion and shape.

Arguments

gi_version: The software version of the Exadata Grid Infrastructure (GI).

shape: The Exadata hardware system model.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of items to return for this request. To get the next page of items, make another request with the token returned in the output.

Default: 10

"nextToken": The token returned from a previous paginated request. Pagination continues from the end of the items returned by the previous request.

Main.Odb.list_tags_for_resource — Function

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Returns information about the tags applied to this resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource to list tags for.

Main.Odb.reboot_db_node — Function

reboot_db_node(cloud_vm_cluster_id, db_node_id)
reboot_db_node(cloud_vm_cluster_id, db_node_id, params::Dict{String,<:Any})

Reboots the specified DB node in a VM cluster.

Arguments

cloud_vm_cluster_id: The unique identifier of the VM cluster that contains the DB node to reboot.

db_node_id: The unique identifier of the DB node to reboot.

Main.Odb.start_db_node — Function

start_db_node(cloud_vm_cluster_id, db_node_id)
start_db_node(cloud_vm_cluster_id, db_node_id, params::Dict{String,<:Any})

Starts the specified DB node in a VM cluster.

Arguments

cloud_vm_cluster_id: The unique identifier of the VM cluster that contains the DB node to start.

db_node_id: The unique identifier of the DB node to start.

Main.Odb.stop_db_node — Function

stop_db_node(cloud_vm_cluster_id, db_node_id)
stop_db_node(cloud_vm_cluster_id, db_node_id, params::Dict{String,<:Any})

Stops the specified DB node in a VM cluster.

Arguments

cloud_vm_cluster_id: The unique identifier of the VM cluster that contains the DB node to stop.

db_node_id: The unique identifier of the DB node to stop.

Main.Odb.tag_resource — Function

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Applies tags to the specified resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource to apply tags to.

tags: The list of tags to apply to the resource.

Main.Odb.untag_resource — Function

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes tags from the specified resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource to remove tags from.

tag_keys: The names (keys) of the tags to remove from the resource.

Main.Odb.update_cloud_exadata_infrastructure — Function

update_cloud_exadata_infrastructure(cloud_exadata_infrastructure_id)
update_cloud_exadata_infrastructure(cloud_exadata_infrastructure_id, params::Dict{String,<:Any})

Updates the properties of an Exadata infrastructure resource.

Arguments

cloud_exadata_infrastructure_id: The unique identifier of the Exadata infrastructure to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maintenanceWindow":

Main.Odb.update_odb_network — Function

update_odb_network(odb_network_id)
update_odb_network(odb_network_id, params::Dict{String,<:Any})

Updates properties of a specified ODB network.

Arguments

odb_network_id: The unique identifier of the ODB network to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"crossRegionS3RestoreSourcesToDisable": The cross-Region Amazon S3 restore sources to disable for the ODB network.

"crossRegionS3RestoreSourcesToEnable": The cross-Region Amazon S3 restore sources to enable for the ODB network.

"displayName": The new user-friendly name of the ODB network.

"kmsAccess": The Amazon Web Services Key Management Service (KMS) access configuration for the ODB network.

"kmsPolicyDocument": The Amazon Web Services Key Management Service (KMS) policy document that defines permissions for key usage within the ODB network.

"peeredCidrsToBeAdded": The list of CIDR ranges from the peered VPC that allow access to the ODB network.

"peeredCidrsToBeRemoved": The list of CIDR ranges from the peered VPC to remove from the ODB network.

"s3Access": Specifies the updated configuration for Amazon S3 access from the ODB network.

"s3PolicyDocument": Specifies the updated endpoint policy for Amazon S3 access from the ODB network.

"stsAccess": The Amazon Web Services Security Token Service (STS) access configuration for the ODB network.

"stsPolicyDocument": The Amazon Web Services Security Token Service (STS) policy document that defines permissions for token service usage within the ODB network.

"zeroEtlAccess": Specifies the updated configuration for Zero-ETL access from the ODB network.

Main.Odb.update_odb_peering_connection — Function

update_odb_peering_connection(odb_peering_connection_id)
update_odb_peering_connection(odb_peering_connection_id, params::Dict{String,<:Any})

Modifies the settings of an Oracle Database@Amazon Web Services peering connection. You can update the display name and add or remove CIDR blocks from the peering connection.

Arguments

odb_peering_connection_id: The identifier of the Oracle Database@Amazon Web Services peering connection to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"displayName": A new display name for the peering connection.

"peerNetworkCidrsToBeAdded": A list of CIDR blocks to add to the peering connection. These CIDR blocks define the IP address ranges that can communicate through the peering connection. The CIDR blocks must not overlap with existing CIDR blocks in the Oracle Database@Amazon Web Services network.

"peerNetworkCidrsToBeRemoved": A list of CIDR blocks to remove from the peering connection. The CIDR blocks must currently exist in the peering connection.