Opensearchserverless

This page documents function available when using the Opensearchserverless module, created with @service Opensearchserverless.

Index

Main.Opensearchserverless.batch_get_collection

Main.Opensearchserverless.batch_get_effective_lifecycle_policy

Main.Opensearchserverless.batch_get_lifecycle_policy

Main.Opensearchserverless.batch_get_vpc_endpoint

Main.Opensearchserverless.create_access_policy

Main.Opensearchserverless.create_collection

Main.Opensearchserverless.create_lifecycle_policy

Main.Opensearchserverless.create_security_config

Main.Opensearchserverless.create_security_policy

Main.Opensearchserverless.create_vpc_endpoint

Main.Opensearchserverless.delete_access_policy

Main.Opensearchserverless.delete_collection

Main.Opensearchserverless.delete_lifecycle_policy

Main.Opensearchserverless.delete_security_config

Main.Opensearchserverless.delete_security_policy

Main.Opensearchserverless.delete_vpc_endpoint

Main.Opensearchserverless.get_access_policy

Main.Opensearchserverless.get_account_settings

Main.Opensearchserverless.get_policies_stats

Main.Opensearchserverless.get_security_config

Main.Opensearchserverless.get_security_policy

Main.Opensearchserverless.list_access_policies

Main.Opensearchserverless.list_collections

Main.Opensearchserverless.list_lifecycle_policies

Main.Opensearchserverless.list_security_configs

Main.Opensearchserverless.list_security_policies

Main.Opensearchserverless.list_tags_for_resource

Main.Opensearchserverless.list_vpc_endpoints

Main.Opensearchserverless.tag_resource

Main.Opensearchserverless.untag_resource

Main.Opensearchserverless.update_access_policy

Main.Opensearchserverless.update_account_settings

Main.Opensearchserverless.update_collection

Main.Opensearchserverless.update_lifecycle_policy

Main.Opensearchserverless.update_security_config

Main.Opensearchserverless.update_security_policy

Main.Opensearchserverless.update_vpc_endpoint

Documentation

Main.Opensearchserverless.batch_get_collection — Method

batch_get_collection()
batch_get_collection(params::Dict{String,<:Any})

Returns attributes for one or more collections, including the collection endpoint and the OpenSearch Dashboards endpoint. For more information, see Creating and managing Amazon OpenSearch Serverless collections.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ids": A list of collection IDs. You can't provide names and IDs in the same request. The ID is part of the collection endpoint. You can also retrieve it using the ListCollections API.

"names": A list of collection names. You can't provide names and IDs in the same request.

Main.Opensearchserverless.batch_get_effective_lifecycle_policy — Method

batch_get_effective_lifecycle_policy(resource_identifiers)
batch_get_effective_lifecycle_policy(resource_identifiers, params::Dict{String,<:Any})

Returns a list of successful and failed retrievals for the OpenSearch Serverless indexes. For more information, see Viewing data lifecycle policies.

Arguments

resource_identifiers: The unique identifiers of policy types and resource names.

Main.Opensearchserverless.batch_get_lifecycle_policy — Method

batch_get_lifecycle_policy(identifiers)
batch_get_lifecycle_policy(identifiers, params::Dict{String,<:Any})

Returns one or more configured OpenSearch Serverless lifecycle policies. For more information, see Viewing data lifecycle policies.

Arguments

identifiers: The unique identifiers of policy types and policy names.

Main.Opensearchserverless.batch_get_vpc_endpoint — Method

batch_get_vpc_endpoint(ids)
batch_get_vpc_endpoint(ids, params::Dict{String,<:Any})

Returns attributes for one or more VPC endpoints associated with the current account. For more information, see Access Amazon OpenSearch Serverless using an interface endpoint.

Arguments

ids: A list of VPC endpoint identifiers.

Main.Opensearchserverless.create_access_policy — Method

create_access_policy(name, policy, type)
create_access_policy(name, policy, type, params::Dict{String,<:Any})

Creates a data access policy for OpenSearch Serverless. Access policies limit access to collections and the resources within them, and allow a user to access that data irrespective of the access mechanism or network source. For more information, see Data access control for Amazon OpenSearch Serverless.

Arguments

name: The name of the policy.

policy: The JSON policy document to use as the content for the policy.

type: The type of policy.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier to ensure idempotency of the request.

"description": A description of the policy. Typically used to store information about the permissions defined in the policy.

Main.Opensearchserverless.create_collection — Method

create_collection(name)
create_collection(name, params::Dict{String,<:Any})

Creates a new OpenSearch Serverless collection. For more information, see Creating and managing Amazon OpenSearch Serverless collections.

Arguments

name: Name of the collection.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier to ensure idempotency of the request.

"description": Description of the collection.

"standbyReplicas": Indicates whether standby replicas should be used for a collection.

"tags": An arbitrary set of tags (key–value pairs) to associate with the OpenSearch Serverless collection.

"type": The type of collection.

Main.Opensearchserverless.create_lifecycle_policy — Method

create_lifecycle_policy(name, policy, type)
create_lifecycle_policy(name, policy, type, params::Dict{String,<:Any})

Creates a lifecyle policy to be applied to OpenSearch Serverless indexes. Lifecycle policies define the number of days or hours to retain the data on an OpenSearch Serverless index. For more information, see Creating data lifecycle policies.

Arguments

name: The name of the lifecycle policy.

policy: The JSON policy document to use as the content for the lifecycle policy.

type: The type of lifecycle policy.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier to ensure idempotency of the request.

"description": A description of the lifecycle policy.

Main.Opensearchserverless.create_security_config — Method

create_security_config(name, type)
create_security_config(name, type, params::Dict{String,<:Any})

Specifies a security configuration for OpenSearch Serverless. For more information, see SAML authentication for Amazon OpenSearch Serverless.

Arguments

name: The name of the security configuration.

type: The type of security configuration.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier to ensure idempotency of the request.

"description": A description of the security configuration.

"samlOptions": Describes SAML options in in the form of a key-value map. This field is required if you specify saml for the type parameter.

Main.Opensearchserverless.create_security_policy — Method

create_security_policy(name, policy, type)
create_security_policy(name, policy, type, params::Dict{String,<:Any})

Creates a security policy to be used by one or more OpenSearch Serverless collections. Security policies provide access to a collection and its OpenSearch Dashboards endpoint from public networks or specific VPC endpoints. They also allow you to secure a collection with a KMS encryption key. For more information, see Network access for Amazon OpenSearch Serverless and Encryption at rest for Amazon OpenSearch Serverless.

Arguments

name: The name of the policy.

policy: The JSON policy document to use as the content for the new policy.

type: The type of security policy.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier to ensure idempotency of the request.

"description": A description of the policy. Typically used to store information about the permissions defined in the policy.

Main.Opensearchserverless.create_vpc_endpoint — Method

create_vpc_endpoint(name, subnet_ids, vpc_id)
create_vpc_endpoint(name, subnet_ids, vpc_id, params::Dict{String,<:Any})

Creates an OpenSearch Serverless-managed interface VPC endpoint. For more information, see Access Amazon OpenSearch Serverless using an interface endpoint.

Arguments

name: The name of the interface endpoint.

subnet_ids: The ID of one or more subnets from which you'll access OpenSearch Serverless.

vpc_id: The ID of the VPC from which you'll access OpenSearch Serverless.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier to ensure idempotency of the request.

"securityGroupIds": The unique identifiers of the security groups that define the ports, protocols, and sources for inbound traffic that you are authorizing into your endpoint.

Main.Opensearchserverless.delete_access_policy — Method

delete_access_policy(name, type)
delete_access_policy(name, type, params::Dict{String,<:Any})

Deletes an OpenSearch Serverless access policy. For more information, see Data access control for Amazon OpenSearch Serverless.

Arguments

name: The name of the policy to delete.

type: The type of policy.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier to ensure idempotency of the request.

Main.Opensearchserverless.delete_collection — Method

delete_collection(id)
delete_collection(id, params::Dict{String,<:Any})

Deletes an OpenSearch Serverless collection. For more information, see Creating and managing Amazon OpenSearch Serverless collections.

Arguments

id: The unique identifier of the collection. For example, 1iu5usc406kd. The ID is part of the collection endpoint. You can also retrieve it using the ListCollections API.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier to ensure idempotency of the request.

Main.Opensearchserverless.delete_lifecycle_policy — Method

delete_lifecycle_policy(name, type)
delete_lifecycle_policy(name, type, params::Dict{String,<:Any})

Deletes an OpenSearch Serverless lifecycle policy. For more information, see Deleting data lifecycle policies.

Arguments

name: The name of the policy to delete.

type: The type of lifecycle policy.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier to ensure idempotency of the request.

Main.Opensearchserverless.delete_security_config — Method

delete_security_config(id)
delete_security_config(id, params::Dict{String,<:Any})

Deletes a security configuration for OpenSearch Serverless. For more information, see SAML authentication for Amazon OpenSearch Serverless.

Arguments

id: The security configuration identifier. For SAML the ID will be saml/&lt;accountId&gt;/&lt;idpProviderName&gt;. For example, saml/123456789123/OKTADev.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier to ensure idempotency of the request.

Main.Opensearchserverless.delete_security_policy — Method

delete_security_policy(name, type)
delete_security_policy(name, type, params::Dict{String,<:Any})

Deletes an OpenSearch Serverless security policy.

Arguments

name: The name of the policy to delete.

type: The type of policy.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier to ensure idempotency of the request.

Main.Opensearchserverless.delete_vpc_endpoint — Method

delete_vpc_endpoint(id)
delete_vpc_endpoint(id, params::Dict{String,<:Any})

Deletes an OpenSearch Serverless-managed interface endpoint. For more information, see Access Amazon OpenSearch Serverless using an interface endpoint.

Arguments

id: The VPC endpoint identifier.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier to ensure idempotency of the request.

Main.Opensearchserverless.get_access_policy — Method

get_access_policy(name, type)
get_access_policy(name, type, params::Dict{String,<:Any})

Returns an OpenSearch Serverless access policy. For more information, see Data access control for Amazon OpenSearch Serverless.

Arguments

name: The name of the access policy.

type: Tye type of policy. Currently, the only supported value is data.

Main.Opensearchserverless.get_account_settings — Method

get_account_settings()
get_account_settings(params::Dict{String,<:Any})

Returns account-level settings related to OpenSearch Serverless.

Main.Opensearchserverless.get_policies_stats — Method

get_policies_stats()
get_policies_stats(params::Dict{String,<:Any})

Returns statistical information about your OpenSearch Serverless access policies, security configurations, and security policies.

Main.Opensearchserverless.get_security_config — Method

get_security_config(id)
get_security_config(id, params::Dict{String,<:Any})

Returns information about an OpenSearch Serverless security configuration. For more information, see SAML authentication for Amazon OpenSearch Serverless.

Arguments

id: The unique identifier of the security configuration.

Main.Opensearchserverless.get_security_policy — Method

get_security_policy(name, type)
get_security_policy(name, type, params::Dict{String,<:Any})

Returns information about a configured OpenSearch Serverless security policy. For more information, see Network access for Amazon OpenSearch Serverless and Encryption at rest for Amazon OpenSearch Serverless.

Arguments

name: The name of the security policy.

type: The type of security policy.

Main.Opensearchserverless.list_access_policies — Method

list_access_policies(type)
list_access_policies(type, params::Dict{String,<:Any})

Returns information about a list of OpenSearch Serverless access policies.

Arguments

type: The type of access policy.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": An optional parameter that specifies the maximum number of results to return. You can use nextToken to get the next page of results. The default is 20.

"nextToken": If your initial ListAccessPolicies operation returns a nextToken, you can include the returned nextToken in subsequent ListAccessPolicies operations, which returns results in the next page.

"resource": Resource filters (can be collections or indexes) that policies can apply to.

Main.Opensearchserverless.list_collections — Method

list_collections()
list_collections(params::Dict{String,<:Any})

Lists all OpenSearch Serverless collections. For more information, see Creating and managing Amazon OpenSearch Serverless collections. Make sure to include an empty request body {} if you don't include any collection filters in the request.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"collectionFilters": A list of filter names and values that you can use for requests.

"maxResults": The maximum number of results to return. Default is 20. You can use nextToken to get the next page of results.

"nextToken": If your initial ListCollections operation returns a nextToken, you can include the returned nextToken in subsequent ListCollections operations, which returns results in the next page.

Main.Opensearchserverless.list_lifecycle_policies — Method

list_lifecycle_policies(type)
list_lifecycle_policies(type, params::Dict{String,<:Any})

Returns a list of OpenSearch Serverless lifecycle policies. For more information, see Viewing data lifecycle policies.

Arguments

type: The type of lifecycle policy.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": An optional parameter that specifies the maximum number of results to return. You can use use nextToken to get the next page of results. The default is 10.

"nextToken": If your initial ListLifecyclePolicies operation returns a nextToken, you can include the returned nextToken in subsequent ListLifecyclePolicies operations, which returns results in the next page.

"resources": Resource filters that policies can apply to. Currently, the only supported resource type is index.

Main.Opensearchserverless.list_security_configs — Method

list_security_configs(type)
list_security_configs(type, params::Dict{String,<:Any})

Returns information about configured OpenSearch Serverless security configurations. For more information, see SAML authentication for Amazon OpenSearch Serverless.

Arguments

type: The type of security configuration.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": An optional parameter that specifies the maximum number of results to return. You can use nextToken to get the next page of results. The default is 20.

"nextToken": If your initial ListSecurityConfigs operation returns a nextToken, you can include the returned nextToken in subsequent ListSecurityConfigs operations, which returns results in the next page.

Main.Opensearchserverless.list_security_policies — Method

list_security_policies(type)
list_security_policies(type, params::Dict{String,<:Any})

Returns information about configured OpenSearch Serverless security policies.

Arguments

type: The type of policy.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": An optional parameter that specifies the maximum number of results to return. You can use nextToken to get the next page of results. The default is 20.

"nextToken": If your initial ListSecurityPolicies operation returns a nextToken, you can include the returned nextToken in subsequent ListSecurityPolicies operations, which returns results in the next page.

"resource": Resource filters (can be collection or indexes) that policies can apply to.

Main.Opensearchserverless.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Returns the tags for an OpenSearch Serverless resource. For more information, see Tagging Amazon OpenSearch Serverless collections.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource. The resource must be active (not in the DELETING state), and must be owned by the account ID included in the request.

Main.Opensearchserverless.list_vpc_endpoints — Method

list_vpc_endpoints()
list_vpc_endpoints(params::Dict{String,<:Any})

Returns the OpenSearch Serverless-managed interface VPC endpoints associated with the current account. For more information, see Access Amazon OpenSearch Serverless using an interface endpoint.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": An optional parameter that specifies the maximum number of results to return. You can use nextToken to get the next page of results. The default is 20.

"nextToken": If your initial ListVpcEndpoints operation returns a nextToken, you can include the returned nextToken in subsequent ListVpcEndpoints operations, which returns results in the next page.

"vpcEndpointFilters": Filter the results according to the current status of the VPC endpoint. Possible statuses are CREATING, DELETING, UPDATING, ACTIVE, and FAILED.

Main.Opensearchserverless.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Associates tags with an OpenSearch Serverless resource. For more information, see Tagging Amazon OpenSearch Serverless collections.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource. The resource must be active (not in the DELETING state), and must be owned by the account ID included in the request.

tags: A list of tags (key-value pairs) to add to the resource. All tag keys in the request must be unique.

Main.Opensearchserverless.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes a tag or set of tags from an OpenSearch Serverless resource. For more information, see Tagging Amazon OpenSearch Serverless collections.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource to remove tags from. The resource must be active (not in the DELETING state), and must be owned by the account ID included in the request.

tag_keys: The tag or set of tags to remove from the resource. All tag keys in the request must be unique.

Main.Opensearchserverless.update_access_policy — Method

update_access_policy(name, policy_version, type)
update_access_policy(name, policy_version, type, params::Dict{String,<:Any})

Updates an OpenSearch Serverless access policy. For more information, see Data access control for Amazon OpenSearch Serverless.

Arguments

name: The name of the policy.

policy_version: The version of the policy being updated.

type: The type of policy.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier to ensure idempotency of the request.

"description": A description of the policy. Typically used to store information about the permissions defined in the policy.

"policy": The JSON policy document to use as the content for the policy.

Main.Opensearchserverless.update_account_settings — Method

update_account_settings()
update_account_settings(params::Dict{String,<:Any})

Update the OpenSearch Serverless settings for the current Amazon Web Services account. For more information, see Managing capacity limits for Amazon OpenSearch Serverless.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"capacityLimits":

Main.Opensearchserverless.update_collection — Method

update_collection(id)
update_collection(id, params::Dict{String,<:Any})

Updates an OpenSearch Serverless collection.

Arguments

id: The unique identifier of the collection.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier to ensure idempotency of the request.

"description": A description of the collection.

Main.Opensearchserverless.update_lifecycle_policy — Method

update_lifecycle_policy(name, policy_version, type)
update_lifecycle_policy(name, policy_version, type, params::Dict{String,<:Any})

Updates an OpenSearch Serverless access policy. For more information, see Updating data lifecycle policies.

Arguments

name: The name of the policy.

policy_version: The version of the policy being updated.

type: The type of lifecycle policy.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier to ensure idempotency of the request.

"description": A description of the lifecycle policy.

"policy": The JSON policy document to use as the content for the lifecycle policy.

Main.Opensearchserverless.update_security_config — Method

update_security_config(config_version, id)
update_security_config(config_version, id, params::Dict{String,<:Any})

Updates a security configuration for OpenSearch Serverless. For more information, see SAML authentication for Amazon OpenSearch Serverless.

Arguments

config_version: The version of the security configuration to be updated. You can find the most recent version of a security configuration using the GetSecurityPolicy command.

id: The security configuration identifier. For SAML the ID will be saml/&lt;accountId&gt;/&lt;idpProviderName&gt;. For example, saml/123456789123/OKTADev.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier to ensure idempotency of the request.

"description": A description of the security configuration.

"samlOptions": SAML options in in the form of a key-value map.

Main.Opensearchserverless.update_security_policy — Method

update_security_policy(name, policy_version, type)
update_security_policy(name, policy_version, type, params::Dict{String,<:Any})

Updates an OpenSearch Serverless security policy. For more information, see Network access for Amazon OpenSearch Serverless and Encryption at rest for Amazon OpenSearch Serverless.

Arguments

name: The name of the policy.

policy_version: The version of the policy being updated.

type: The type of access policy.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier to ensure idempotency of the request.

"description": A description of the policy. Typically used to store information about the permissions defined in the policy.

"policy": The JSON policy document to use as the content for the new policy.

Main.Opensearchserverless.update_vpc_endpoint — Method

update_vpc_endpoint(id)
update_vpc_endpoint(id, params::Dict{String,<:Any})

Updates an OpenSearch Serverless-managed interface endpoint. For more information, see Access Amazon OpenSearch Serverless using an interface endpoint.

Arguments

id: The unique identifier of the interface endpoint to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"addSecurityGroupIds": The unique identifiers of the security groups to add to the endpoint. Security groups define the ports, protocols, and sources for inbound traffic that you are authorizing into your endpoint.

"addSubnetIds": The ID of one or more subnets to add to the endpoint.

"clientToken": Unique, case-sensitive identifier to ensure idempotency of the request.

"removeSecurityGroupIds": The unique identifiers of the security groups to remove from the endpoint.

"removeSubnetIds": The unique identifiers of the subnets to remove from the endpoint.