Opsworkscm

This page documents function available when using the Opsworkscm module, created with @service Opsworkscm.

Index

Main.Opsworkscm.associate_node

Main.Opsworkscm.create_backup

Main.Opsworkscm.create_server

Main.Opsworkscm.delete_backup

Main.Opsworkscm.delete_server

Main.Opsworkscm.describe_account_attributes

Main.Opsworkscm.describe_backups

Main.Opsworkscm.describe_events

Main.Opsworkscm.describe_node_association_status

Main.Opsworkscm.describe_servers

Main.Opsworkscm.disassociate_node

Main.Opsworkscm.export_server_engine_attribute

Main.Opsworkscm.list_tags_for_resource

Main.Opsworkscm.restore_server

Main.Opsworkscm.start_maintenance

Main.Opsworkscm.tag_resource

Main.Opsworkscm.untag_resource

Main.Opsworkscm.update_server

Main.Opsworkscm.update_server_engine_attributes

Documentation

Main.Opsworkscm.associate_node — Method

associate_node(engine_attributes, node_name, server_name)
associate_node(engine_attributes, node_name, server_name, params::Dict{String,<:Any})

Associates a new node with the server. For more information about how to disassociate a node, see DisassociateNode. On a Chef server: This command is an alternative to knife bootstrap. Example (Chef): aws opsworks-cm associate-node –server-name MyServer –node-name MyManagedNode –engine-attributes "Name=CHEFORGANIZATION,Value=default" "Name=CHEFNODEPUBLICKEY,Value=public-key-pem" On a Puppet server, this command is an alternative to the puppet cert sign command that signs a Puppet node CSR. Example (Puppet): aws opsworks-cm associate-node –server-name MyServer –node-name MyManagedNode –engine-attributes "Name=PUPPETNODECSR,Value=csr-pem" A node can can only be associated with servers that are in a HEALTHY state. Otherwise, an InvalidStateException is thrown. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid. The AssociateNode API call can be integrated into Auto Scaling configurations, AWS Cloudformation templates, or the user data of a server's instance.

Arguments

engine_attributes: Engine attributes used for associating the node. Attributes accepted in a AssociateNode request for Chef CHEFORGANIZATION: The Chef organization with which the node is associated. By default only one organization named default can exist. CHEFNODEPUBLICKEY: A PEM-formatted public key. This key is required for the chef-client agent to access the Chef API. Attributes accepted in a AssociateNode request for Puppet PUPPETNODECSR: A PEM-formatted certificate-signing request (CSR) that is created by the node.

node_name: The name of the node.

server_name: The name of the server with which to associate the node.

Main.Opsworkscm.create_backup — Method

create_backup(server_name)
create_backup(server_name, params::Dict{String,<:Any})

Creates an application-level backup of a server. While the server is in the BACKING_UP state, the server cannot be changed, and no additional backup can be created. Backups can be created for servers in RUNNING, HEALTHY, and UNHEALTHY states. By default, you can create a maximum of 50 manual backups. This operation is asynchronous. A LimitExceededException is thrown when the maximum number of manual backups is reached. An InvalidStateException is thrown when the server is not in any of the following states: RUNNING, HEALTHY, or UNHEALTHY. A ResourceNotFoundException is thrown when the server is not found. A ValidationException is thrown when parameters of the request are not valid.

Arguments

server_name: The name of the server that you want to back up.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A user-defined description of the backup.

"Tags": A map that contains tag keys and tag values to attach to an AWS OpsWorks-CM server backup. The key cannot be empty. The key can be a maximum of 127 characters, and can contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : / The value can be a maximum 255 characters, and contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : / Leading and trailing white spaces are trimmed from both the key and value. A maximum of 50 user-applied tags is allowed for tag-supported AWS OpsWorks-CM resources.

Main.Opsworkscm.create_server — Method

create_server(engine, instance_profile_arn, instance_type, server_name, service_role_arn)
create_server(engine, instance_profile_arn, instance_type, server_name, service_role_arn, params::Dict{String,<:Any})

Creates and immedately starts a new server. The server is ready to use when it is in the HEALTHY state. By default, you can create a maximum of 10 servers. This operation is asynchronous. A LimitExceededException is thrown when you have created the maximum number of servers (10). A ResourceAlreadyExistsException is thrown when a server with the same name already exists in the account. A ResourceNotFoundException is thrown when you specify a backup ID that is not valid or is for a backup that does not exist. A ValidationException is thrown when parameters of the request are not valid. If you do not specify a security group by adding the SecurityGroupIds parameter, AWS OpsWorks creates a new security group. Chef Automate: The default security group opens the Chef server to the world on TCP port

If a KeyName is present, AWS OpsWorks enables SSH access. SSH is also open to the

world on TCP port 22. Puppet Enterprise: The default security group opens TCP ports 22, 443, 4433, 8140, 8142, 8143, and 8170. If a KeyName is present, AWS OpsWorks enables SSH access. SSH is also open to the world on TCP port 22. By default, your server is accessible from any IP address. We recommend that you update your security group rules to allow access from known IP addresses and address ranges only. To edit security group rules, open Security Groups in the navigation pane of the EC2 management console. To specify your own domain for a server, and provide your own self-signed or CA-signed certificate and private key, specify values for CustomDomain, CustomCertificate, and CustomPrivateKey.

Arguments

engine: The configuration management engine to use. Valid values include ChefAutomate and Puppet.

instance_profile_arn: The ARN of the instance profile that your Amazon EC2 instances use. Although the AWS OpsWorks console typically creates the instance profile for you, if you are using API commands instead, run the service-role-creation.yaml AWS CloudFormation template, located at https://s3.amazonaws.com/opsworks-cm-us-east-1-prod-default-assets/misc/opsworks-cm-roles.ya ml. This template creates a CloudFormation stack that includes the instance profile you need.

instance_type: The Amazon EC2 instance type to use. For example, m5.large.

server_name: The name of the server. The server name must be unique within your AWS account, within each region. Server names must start with a letter; then letters, numbers, or hyphens (-) are allowed, up to a maximum of 40 characters.

service_role_arn: The service role that the AWS OpsWorks CM service backend uses to work with your account. Although the AWS OpsWorks management console typically creates the service role for you, if you are using the AWS CLI or API commands, run the service-role-creation.yaml AWS CloudFormation template, located at https://s3.amazonaws.com/opsworks-cm-us-east-1-prod-default-assets/misc/opsworks-cm-roles.ya ml. This template creates a CloudFormation stack that includes the service role and instance profile that you need.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AssociatePublicIpAddress": Associate a public IP address with a server that you are launching. Valid values are true or false. The default value is true.

"BackupId": If you specify this field, AWS OpsWorks CM creates the server by using the backup represented by BackupId.

"BackupRetentionCount": The number of automated backups that you want to keep. Whenever a new backup is created, AWS OpsWorks CM deletes the oldest backups if this number is exceeded. The default value is 1.

"CustomCertificate": A PEM-formatted HTTPS certificate. The value can be be a single, self-signed certificate, or a certificate chain. If you specify a custom certificate, you must also specify values for CustomDomain and CustomPrivateKey. The following are requirements for the CustomCertificate value: You can provide either a self-signed, custom certificate, or the full certificate chain. The certificate must be a valid X509 certificate, or a certificate chain in PEM format. The certificate must be valid at the time of upload. A certificate can't be used before its validity period begins (the certificate's NotBefore date), or after it expires (the certificate's NotAfter date). The certificate’s common name or subject alternative names (SANs), if present, must match the value of CustomDomain. The certificate must match the value of CustomPrivateKey.

"CustomDomain": An optional public endpoint of a server, such as https://aws.my-company.com. To access the server, create a CNAME DNS record in your preferred DNS service that points the custom domain to the endpoint that is generated when the server is created (the value of the CreateServer Endpoint attribute). You cannot access the server by using the generated Endpoint value if the server is using a custom domain. If you specify a custom domain, you must also specify values for CustomCertificate and CustomPrivateKey.

"CustomPrivateKey": A private key in PEM format for connecting to the server by using HTTPS. The private key must not be encrypted; it cannot be protected by a password or passphrase. If you specify a custom private key, you must also specify values for CustomDomain and CustomCertificate.

"DisableAutomatedBackup": Enable or disable scheduled backups. Valid values are true or false. The default value is true.

"EngineAttributes": Optional engine attributes on a specified server. Attributes accepted in a Chef createServer request: CHEFAUTOMATEPIVOTALKEY: A base64-encoded RSA public key. The corresponding private key is required to access the Chef API. When no CHEFAUTOMATEPIVOTALKEY is set, a private key is generated and returned in the response. CHEFAUTOMATEADMINPASSWORD: The password for the administrative user in the Chef Automate web-based dashboard. The password length is a minimum of eight characters, and a maximum of 32. The password can contain letters, numbers, and special characters (!/@#%^&amp;+=). The password must contain at least one lower case letter, one upper case letter, one number, and one special character. When no CHEFAUTOMATEADMINPASSWORD is set, one is generated and returned in the response. Attributes accepted in a Puppet createServer request: PUPPETADMINPASSWORD: To work with the Puppet Enterprise console, a password must use ASCII characters. PUPPETR10KREMOTE: The r10k remote is the URL of your control repository (for example, ssh://git@your.git-repo.com:user/control-repo.git). Specifying an r10k remote opens TCP port 8170. PUPPETR10KPRIVATEKEY: If you are using a private Git repository, add PUPPETR10KPRIVATE_KEY to specify a PEM-encoded private SSH key.

"EngineModel": The engine model of the server. Valid values in this release include Monolithic for Puppet and Single for Chef.

"EngineVersion": The major release version of the engine that you want to use. For a Chef server, the valid value for EngineVersion is currently 2. For a Puppet server, valid values are 2019 or 2017.

"KeyPair": The Amazon EC2 key pair to set for the instance. This parameter is optional; if desired, you may specify this parameter to connect to your instances by using SSH.

"PreferredBackupWindow": The start time for a one-hour period during which AWS OpsWorks CM backs up application-level data on your server if automated backups are enabled. Valid values must be specified in one of the following formats: HH:MM for daily backups DDD:HH:MM for weekly backups MM must be specified as 00. The specified time is in coordinated universal time (UTC). The default value is a random, daily start time. Example: 08:00, which represents a daily start time of 08:00 UTC. Example: Mon:08:00, which represents a start time of every Monday at 08:00 UTC. (8:00 a.m.)

"PreferredMaintenanceWindow": The start time for a one-hour period each week during which AWS OpsWorks CM performs maintenance on the instance. Valid values must be specified in the following format: DDD:HH:MM. MM must be specified as 00. The specified time is in coordinated universal time (UTC). The default value is a random one-hour period on Tuesday, Wednesday, or Friday. See TimeWindowDefinition for more information. Example: Mon:08:00, which represents a start time of every Monday at 08:00 UTC. (8:00 a.m.)

"SecurityGroupIds": A list of security group IDs to attach to the Amazon EC2 instance. If you add this parameter, the specified security groups must be within the VPC that is specified by SubnetIds. If you do not specify this parameter, AWS OpsWorks CM creates one new security group that uses TCP ports 22 and 443, open to 0.0.0.0/0 (everyone).

"SubnetIds": The IDs of subnets in which to launch the server EC2 instance. Amazon EC2-Classic customers: This field is required. All servers must run within a VPC. The VPC must have "Auto Assign Public IP" enabled. EC2-VPC customers: This field is optional. If you do not specify subnet IDs, your EC2 instances are created in a default subnet that is selected by Amazon EC2. If you specify subnet IDs, the VPC must have "Auto Assign Public IP" enabled. For more information about supported Amazon EC2 platforms, see Supported Platforms.

"Tags": A map that contains tag keys and tag values to attach to an AWS OpsWorks for Chef Automate or AWS OpsWorks for Puppet Enterprise server. The key cannot be empty. The key can be a maximum of 127 characters, and can contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : / @ The value can be a maximum 255 characters, and contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : / @ Leading and trailing white spaces are trimmed from both the key and value. A maximum of 50 user-applied tags is allowed for any AWS OpsWorks-CM server.

Main.Opsworkscm.delete_backup — Method

delete_backup(backup_id)
delete_backup(backup_id, params::Dict{String,<:Any})

Deletes a backup. You can delete both manual and automated backups. This operation is asynchronous. An InvalidStateException is thrown when a backup deletion is already in progress. A ResourceNotFoundException is thrown when the backup does not exist. A ValidationException is thrown when parameters of the request are not valid.

Arguments

backup_id: The ID of the backup to delete. Run the DescribeBackups command to get a list of backup IDs. Backup IDs are in the format ServerName-yyyyMMddHHmmssSSS.

Main.Opsworkscm.delete_server — Method

delete_server(server_name)
delete_server(server_name, params::Dict{String,<:Any})

Deletes the server and the underlying AWS CloudFormation stacks (including the server's EC2 instance). When you run this command, the server state is updated to DELETING. After the server is deleted, it is no longer returned by DescribeServer requests. If the AWS CloudFormation stack cannot be deleted, the server cannot be deleted. This operation is asynchronous. An InvalidStateException is thrown when a server deletion is already in progress. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

Arguments

server_name: The ID of the server to delete.

Main.Opsworkscm.describe_account_attributes — Method

describe_account_attributes()
describe_account_attributes(params::Dict{String,<:Any})

Describes your OpsWorks-CM account attributes. This operation is synchronous.

Main.Opsworkscm.describe_backups — Method

describe_backups()
describe_backups(params::Dict{String,<:Any})

Describes backups. The results are ordered by time, with newest backups first. If you do not specify a BackupId or ServerName, the command returns all backups. This operation is synchronous. A ResourceNotFoundException is thrown when the backup does not exist. A ValidationException is raised when parameters of the request are not valid.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"BackupId": Describes a single backup.

"MaxResults": This is not currently implemented for DescribeBackups requests.

"NextToken": This is not currently implemented for DescribeBackups requests.

"ServerName": Returns backups for the server with the specified ServerName.

Main.Opsworkscm.describe_events — Method

describe_events(server_name)
describe_events(server_name, params::Dict{String,<:Any})

Describes events for a specified server. Results are ordered by time, with newest events first. This operation is synchronous. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

Arguments

server_name: The name of the server for which you want to view events.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": To receive a paginated response, use this parameter to specify the maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

"NextToken": NextToken is a string that is returned in some command responses. It indicates that not all entries have been returned, and that you must run at least one more request to get remaining items. To get remaining results, call DescribeEvents again, and assign the token from the previous results as the value of the nextToken parameter. If there are no more results, the response object's nextToken parameter value is null. Setting a nextToken value that was not returned in your previous results causes an InvalidNextTokenException to occur.

Main.Opsworkscm.describe_node_association_status — Method

describe_node_association_status(node_association_status_token, server_name)
describe_node_association_status(node_association_status_token, server_name, params::Dict{String,<:Any})

Returns the current status of an existing association or disassociation request. A ResourceNotFoundException is thrown when no recent association or disassociation request with the specified token is found, or when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

Arguments

node_association_status_token: The token returned in either the AssociateNodeResponse or the DisassociateNodeResponse.

server_name: The name of the server from which to disassociate the node.

Main.Opsworkscm.describe_servers — Method

describe_servers()
describe_servers(params::Dict{String,<:Any})

Lists all configuration management servers that are identified with your account. Only the stored results from Amazon DynamoDB are returned. AWS OpsWorks CM does not query other services. This operation is synchronous. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": This is not currently implemented for DescribeServers requests.

"NextToken": This is not currently implemented for DescribeServers requests.

"ServerName": Describes the server with the specified ServerName.

Main.Opsworkscm.disassociate_node — Method

disassociate_node(node_name, server_name)
disassociate_node(node_name, server_name, params::Dict{String,<:Any})

Disassociates a node from an AWS OpsWorks CM server, and removes the node from the server's managed nodes. After a node is disassociated, the node key pair is no longer valid for accessing the configuration manager's API. For more information about how to associate a node, see AssociateNode. A node can can only be disassociated from a server that is in a HEALTHY state. Otherwise, an InvalidStateException is thrown. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

Arguments

node_name: The name of the client node.

server_name: The name of the server from which to disassociate the node.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"EngineAttributes": Engine attributes that are used for disassociating the node. No attributes are required for Puppet. Attributes required in a DisassociateNode request for Chef CHEF_ORGANIZATION: The Chef organization with which the node was associated. By default only one organization named default can exist.

Main.Opsworkscm.export_server_engine_attribute — Method

export_server_engine_attribute(export_attribute_name, server_name)
export_server_engine_attribute(export_attribute_name, server_name, params::Dict{String,<:Any})

Exports a specified server engine attribute as a base64-encoded string. For example, you can export user data that you can use in EC2 to associate nodes with a server. This operation is synchronous. A ValidationException is raised when parameters of the request are not valid. A ResourceNotFoundException is thrown when the server does not exist. An InvalidStateException is thrown when the server is in any of the following states: CREATING, TERMINATED, FAILED or DELETING.

Arguments

export_attribute_name: The name of the export attribute. Currently, the supported export attribute is Userdata. This exports a user data script that includes parameters and values provided in the InputAttributes list.

server_name: The name of the server from which you are exporting the attribute.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"InputAttributes": The list of engine attributes. The list type is EngineAttribute. An EngineAttribute list item is a pair that includes an attribute name and its value. For the Userdata ExportAttributeName, the following are supported engine attribute names. RunList In Chef, a list of roles or recipes that are run in the specified order. In Puppet, this parameter is ignored. OrganizationName In Chef, an organization name. AWS OpsWorks for Chef Automate always creates the organization default. In Puppet, this parameter is ignored. NodeEnvironment In Chef, a node environment (for example, development, staging, or one-box). In Puppet, this parameter is ignored. NodeClientVersion In Chef, the version of the Chef engine (three numbers separated by dots, such as 13.8.5). If this attribute is empty, OpsWorks for Chef Automate uses the most current version. In Puppet, this parameter is ignored.

Main.Opsworkscm.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Returns a list of tags that are applied to the specified AWS OpsWorks for Chef Automate or AWS OpsWorks for Puppet Enterprise servers or backups.

Arguments

resource_arn: The Amazon Resource Number (ARN) of an AWS OpsWorks for Chef Automate or AWS OpsWorks for Puppet Enterprise server for which you want to show applied tags. For example, arn:aws:opsworks-cm:us-west-2:123456789012:server/test-owcm-server/EXAMPLE-66b0-4196-8274-d1 a2bEXAMPLE.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": To receive a paginated response, use this parameter to specify the maximum number of results to be returned with a single call. If the number of available results exceeds this maximum, the response includes a NextToken value that you can assign to the NextToken request parameter to get the next set of results.

"NextToken": NextToken is a string that is returned in some command responses. It indicates that not all entries have been returned, and that you must run at least one more request to get remaining items. To get remaining results, call ListTagsForResource again, and assign the token from the previous results as the value of the nextToken parameter. If there are no more results, the response object's nextToken parameter value is null. Setting a nextToken value that was not returned in your previous results causes an InvalidNextTokenException to occur.

Main.Opsworkscm.restore_server — Method

restore_server(backup_id, server_name)
restore_server(backup_id, server_name, params::Dict{String,<:Any})

Restores a backup to a server that is in a CONNECTION_LOST, HEALTHY, RUNNING, UNHEALTHY, or TERMINATED state. When you run RestoreServer, the server's EC2 instance is deleted, and a new EC2 instance is configured. RestoreServer maintains the existing server endpoint, so configuration management of the server's client devices (nodes) should continue to work. Restoring from a backup is performed by creating a new EC2 instance. If restoration is successful, and the server is in a HEALTHY state, AWS OpsWorks CM switches traffic over to the new instance. After restoration is finished, the old EC2 instance is maintained in a Running or Stopped state, but is eventually terminated. This operation is asynchronous. An InvalidStateException is thrown when the server is not in a valid state. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

Arguments

backup_id: The ID of the backup that you want to use to restore a server.

server_name: The name of the server that you want to restore.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"InstanceType": The type of instance to restore. Valid values must be specified in the following format: ^([cm][34]|t2).* For example, m5.large. Valid values are m5.large, r5.xlarge, and r5.2xlarge. If you do not specify this parameter, RestoreServer uses the instance type from the specified backup.

"KeyPair": The name of the key pair to set on the new EC2 instance. This can be helpful if the administrator no longer has the SSH key.

Main.Opsworkscm.start_maintenance — Method

start_maintenance(server_name)
start_maintenance(server_name, params::Dict{String,<:Any})

Manually starts server maintenance. This command can be useful if an earlier maintenance attempt failed, and the underlying cause of maintenance failure has been resolved. The server is in an UNDER_MAINTENANCE state while maintenance is in progress. Maintenance can only be started on servers in HEALTHY and UNHEALTHY states. Otherwise, an InvalidStateException is thrown. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

Arguments

server_name: The name of the server on which to run maintenance.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"EngineAttributes": Engine attributes that are specific to the server on which you want to run maintenance. Attributes accepted in a StartMaintenance request for Chef CHEFMAJORUPGRADE: If a Chef Automate server is eligible for upgrade to Chef Automate 2, add this engine attribute to a StartMaintenance request and set the value to true to upgrade the server to Chef Automate 2. For more information, see Upgrade an AWS OpsWorks for Chef Automate Server to Chef Automate 2.

Main.Opsworkscm.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Applies tags to an AWS OpsWorks for Chef Automate or AWS OpsWorks for Puppet Enterprise server, or to server backups.

Arguments

resource_arn: The Amazon Resource Number (ARN) of a resource to which you want to apply tags. For example, arn:aws:opsworks-cm:us-west-2:123456789012:server/test-owcm-server/EXAMPLE-66b0-4196-8274-d1 a2bEXAMPLE.

tags: A map that contains tag keys and tag values to attach to AWS OpsWorks-CM servers or backups. The key cannot be empty. The key can be a maximum of 127 characters, and can contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : / The value can be a maximum 255 characters, and contain only Unicode letters, numbers, or separators, or the following special characters: + - = . _ : / Leading and trailing white spaces are trimmed from both the key and value. A maximum of 50 user-applied tags is allowed for any AWS OpsWorks-CM server or backup.

Main.Opsworkscm.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes specified tags from an AWS OpsWorks-CM server or backup.

Arguments

resource_arn: The Amazon Resource Number (ARN) of a resource from which you want to remove tags. For example, arn:aws:opsworks-cm:us-west-2:123456789012:server/test-owcm-server/EXAMPLE-66b0-4196-8274-d1 a2bEXAMPLE.

tag_keys: The keys of tags that you want to remove.

Main.Opsworkscm.update_server — Method

update_server(server_name)
update_server(server_name, params::Dict{String,<:Any})

Updates settings for a server. This operation is synchronous.

Arguments

server_name: The name of the server to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"BackupRetentionCount": Sets the number of automated backups that you want to keep.

"DisableAutomatedBackup": Setting DisableAutomatedBackup to true disables automated or scheduled backups. Automated backups are enabled by default.

"PreferredBackupWindow":

"PreferredMaintenanceWindow":

Main.Opsworkscm.update_server_engine_attributes — Method

update_server_engine_attributes(attribute_name, server_name)
update_server_engine_attributes(attribute_name, server_name, params::Dict{String,<:Any})

Updates engine-specific attributes on a specified server. The server enters the MODIFYING state when this operation is in progress. Only one update can occur at a time. You can use this command to reset a Chef server's public key (CHEFPIVOTALKEY) or a Puppet server's admin password (PUPPETADMINPASSWORD). This operation is asynchronous. This operation can only be called for servers in HEALTHY or UNHEALTHY states. Otherwise, an InvalidStateException is raised. A ResourceNotFoundException is thrown when the server does not exist. A ValidationException is raised when parameters of the request are not valid.

Arguments

attribute_name: The name of the engine attribute to update.

server_name: The name of the server to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AttributeValue": The value to set for the attribute.