Osis

This page documents function available when using the Osis module, created with @service Osis.

Index

Main.Osis.create_pipeline

Main.Osis.delete_pipeline

Main.Osis.get_pipeline

Main.Osis.get_pipeline_blueprint

Main.Osis.get_pipeline_change_progress

Main.Osis.list_pipeline_blueprints

Main.Osis.list_pipelines

Main.Osis.list_tags_for_resource

Main.Osis.start_pipeline

Main.Osis.stop_pipeline

Main.Osis.tag_resource

Main.Osis.untag_resource

Main.Osis.update_pipeline

Main.Osis.validate_pipeline

Documentation

Main.Osis.create_pipeline — Method

create_pipeline(max_units, min_units, pipeline_configuration_body, pipeline_name)
create_pipeline(max_units, min_units, pipeline_configuration_body, pipeline_name, params::Dict{String,<:Any})

Creates an OpenSearch Ingestion pipeline. For more information, see Creating Amazon OpenSearch Ingestion pipelines.

Arguments

max_units: The maximum pipeline capacity, in Ingestion Compute Units (ICUs).

min_units: The minimum pipeline capacity, in Ingestion Compute Units (ICUs).

pipeline_configuration_body: The pipeline configuration in YAML format. The command accepts the pipeline configuration as a string or within a .yaml file. If you provide the configuration as a string, each new line must be escaped with n.

pipeline_name: The name of the OpenSearch Ingestion pipeline to create. Pipeline names are unique across the pipelines owned by an account within an Amazon Web Services Region.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"BufferOptions": Key-value pairs to configure persistent buffering for the pipeline.

"EncryptionAtRestOptions": Key-value pairs to configure encryption for data that is written to a persistent buffer.

"LogPublishingOptions": Key-value pairs to configure log publishing.

"Tags": List of tags to add to the pipeline upon creation.

"VpcOptions": Container for the values required to configure VPC access for the pipeline. If you don't specify these values, OpenSearch Ingestion creates the pipeline with a public endpoint.

Main.Osis.delete_pipeline — Method

delete_pipeline(pipeline_name)
delete_pipeline(pipeline_name, params::Dict{String,<:Any})

Deletes an OpenSearch Ingestion pipeline. For more information, see Deleting Amazon OpenSearch Ingestion pipelines.

Arguments

pipeline_name: The name of the pipeline to delete.

Main.Osis.get_pipeline — Method

get_pipeline(pipeline_name)
get_pipeline(pipeline_name, params::Dict{String,<:Any})

Retrieves information about an OpenSearch Ingestion pipeline.

Arguments

pipeline_name: The name of the pipeline.

Main.Osis.get_pipeline_blueprint — Method

get_pipeline_blueprint(blueprint_name)
get_pipeline_blueprint(blueprint_name, params::Dict{String,<:Any})

Retrieves information about a specific blueprint for OpenSearch Ingestion. Blueprints are templates for the configuration needed for a CreatePipeline request. For more information, see Using blueprints to create a pipeline.

Arguments

blueprint_name: The name of the blueprint to retrieve.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"format": The format format of the blueprint to retrieve.

Main.Osis.get_pipeline_change_progress — Method

get_pipeline_change_progress(pipeline_name)
get_pipeline_change_progress(pipeline_name, params::Dict{String,<:Any})

Returns progress information for the current change happening on an OpenSearch Ingestion pipeline. Currently, this operation only returns information when a pipeline is being created. For more information, see Tracking the status of pipeline creation.

Arguments

pipeline_name: The name of the pipeline.

Main.Osis.list_pipeline_blueprints — Method

list_pipeline_blueprints()
list_pipeline_blueprints(params::Dict{String,<:Any})

Retrieves a list of all available blueprints for Data Prepper. For more information, see Using blueprints to create a pipeline.

Main.Osis.list_pipelines — Method

list_pipelines()
list_pipelines(params::Dict{String,<:Any})

Lists all OpenSearch Ingestion pipelines in the current Amazon Web Services account and Region. For more information, see Viewing Amazon OpenSearch Ingestion pipelines.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": An optional parameter that specifies the maximum number of results to return. You can use nextToken to get the next page of results.

"nextToken": If your initial ListPipelines operation returns a nextToken, you can include the returned nextToken in subsequent ListPipelines operations, which returns results in the next page.

Main.Osis.list_tags_for_resource — Method

list_tags_for_resource(arn)
list_tags_for_resource(arn, params::Dict{String,<:Any})

Lists all resource tags associated with an OpenSearch Ingestion pipeline. For more information, see Tagging Amazon OpenSearch Ingestion pipelines.

Arguments

arn: The Amazon Resource Name (ARN) of the pipeline to retrieve tags for.

Main.Osis.start_pipeline — Method

start_pipeline(pipeline_name)
start_pipeline(pipeline_name, params::Dict{String,<:Any})

Starts an OpenSearch Ingestion pipeline. For more information, see Starting an OpenSearch Ingestion pipeline.

Arguments

pipeline_name: The name of the pipeline to start.

Main.Osis.stop_pipeline — Method

stop_pipeline(pipeline_name)
stop_pipeline(pipeline_name, params::Dict{String,<:Any})

Stops an OpenSearch Ingestion pipeline. For more information, see Stopping an OpenSearch Ingestion pipeline.

Arguments

pipeline_name: The name of the pipeline to stop.

Main.Osis.tag_resource — Method

tag_resource(tags, arn)
tag_resource(tags, arn, params::Dict{String,<:Any})

Tags an OpenSearch Ingestion pipeline. For more information, see Tagging Amazon OpenSearch Ingestion pipelines.

Arguments

tags: The list of key-value tags to add to the pipeline.

arn: The Amazon Resource Name (ARN) of the pipeline to tag.

Main.Osis.untag_resource — Method

untag_resource(tag_keys, arn)
untag_resource(tag_keys, arn, params::Dict{String,<:Any})

Removes one or more tags from an OpenSearch Ingestion pipeline. For more information, see Tagging Amazon OpenSearch Ingestion pipelines.

Arguments

tag_keys: The tag keys to remove.

arn: The Amazon Resource Name (ARN) of the pipeline to remove tags from.

Main.Osis.update_pipeline — Method

update_pipeline(pipeline_name)
update_pipeline(pipeline_name, params::Dict{String,<:Any})

Updates an OpenSearch Ingestion pipeline. For more information, see Updating Amazon OpenSearch Ingestion pipelines.

Arguments

pipeline_name: The name of the pipeline to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"BufferOptions": Key-value pairs to configure persistent buffering for the pipeline.

"EncryptionAtRestOptions": Key-value pairs to configure encryption for data that is written to a persistent buffer.

"LogPublishingOptions": Key-value pairs to configure log publishing.

"MaxUnits": The maximum pipeline capacity, in Ingestion Compute Units (ICUs)

"MinUnits": The minimum pipeline capacity, in Ingestion Compute Units (ICUs).

"PipelineConfigurationBody": The pipeline configuration in YAML format. The command accepts the pipeline configuration as a string or within a .yaml file. If you provide the configuration as a string, each new line must be escaped with n.

Main.Osis.validate_pipeline — Method

validate_pipeline(pipeline_configuration_body)
validate_pipeline(pipeline_configuration_body, params::Dict{String,<:Any})

Checks whether an OpenSearch Ingestion pipeline configuration is valid prior to creation. For more information, see Creating Amazon OpenSearch Ingestion pipelines.

Arguments

pipeline_configuration_body: The pipeline configuration in YAML format. The command accepts the pipeline configuration as a string or within a .yaml file. If you provide the configuration as a string, each new line must be escaped with n.