Outposts

This page documents function available when using the Outposts module, created with @service Outposts.

Index

Main.Outposts.cancel_order

Main.Outposts.create_order

Main.Outposts.create_outpost

Main.Outposts.create_site

Main.Outposts.delete_outpost

Main.Outposts.delete_site

Main.Outposts.get_catalog_item

Main.Outposts.get_connection

Main.Outposts.get_order

Main.Outposts.get_outpost

Main.Outposts.get_outpost_instance_types

Main.Outposts.get_site

Main.Outposts.get_site_address

Main.Outposts.list_assets

Main.Outposts.list_catalog_items

Main.Outposts.list_orders

Main.Outposts.list_outposts

Main.Outposts.list_sites

Main.Outposts.list_tags_for_resource

Main.Outposts.start_connection

Main.Outposts.tag_resource

Main.Outposts.untag_resource

Main.Outposts.update_outpost

Main.Outposts.update_site

Main.Outposts.update_site_address

Main.Outposts.update_site_rack_physical_properties

Documentation

Main.Outposts.cancel_order — Method

cancel_order(order_id)
cancel_order(order_id, params::Dict{String,<:Any})

Cancels the specified order for an Outpost.

Arguments

order_id: The ID of the order.

Main.Outposts.create_order — Method

create_order(line_items, outpost_identifier, payment_option)
create_order(line_items, outpost_identifier, payment_option, params::Dict{String,<:Any})

Creates an order for an Outpost.

Arguments

line_items: The line items that make up the order.

outpost_identifier: The ID or the Amazon Resource Name (ARN) of the Outpost.

payment_option: The payment option.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"PaymentTerm": The payment terms.

Main.Outposts.create_outpost — Method

create_outpost(name, site_id)
create_outpost(name, site_id, params::Dict{String,<:Any})

Creates an Outpost. You can specify either an Availability one or an AZ ID.

Arguments

name:

site_id: The ID or the Amazon Resource Name (ARN) of the site.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AvailabilityZone":

"AvailabilityZoneId":

"Description":

"SupportedHardwareType": The type of hardware for this Outpost.

"Tags": The tags to apply to the Outpost.

Main.Outposts.create_site — Method

create_site(name)
create_site(name, params::Dict{String,<:Any})

Creates a site for an Outpost.

Arguments

name:

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description":

"Notes": Additional information that you provide about site access requirements, electrician scheduling, personal protective equipment, or regulation of equipment materials that could affect your installation process.

"OperatingAddress": The location to install and power on the hardware. This address might be different from the shipping address.

"RackPhysicalProperties": Information about the physical and logistical details for the rack at this site. For more information about hardware requirements for racks, see Network readiness checklist in the Amazon Web Services Outposts User Guide.

"ShippingAddress": The location to ship the hardware. This address might be different from the operating address.

"Tags": The tags to apply to a site.

Main.Outposts.delete_outpost — Method

delete_outpost(outpost_id)
delete_outpost(outpost_id, params::Dict{String,<:Any})

Deletes the specified Outpost.

Arguments

outpost_id: The ID or the Amazon Resource Name (ARN) of the Outpost.

Main.Outposts.delete_site — Method

delete_site(site_id)
delete_site(site_id, params::Dict{String,<:Any})

Deletes the specified site.

Arguments

site_id: The ID or the Amazon Resource Name (ARN) of the site.

Main.Outposts.get_catalog_item — Method

get_catalog_item(catalog_item_id)
get_catalog_item(catalog_item_id, params::Dict{String,<:Any})

Gets information about the specified catalog item.

Arguments

catalog_item_id: The ID of the catalog item.

Main.Outposts.get_connection — Method

get_connection(connection_id)
get_connection(connection_id, params::Dict{String,<:Any})

Amazon Web Services uses this action to install Outpost servers. Gets information about the specified connection. Use CloudTrail to monitor this action or Amazon Web Services managed policy for Amazon Web Services Outposts to secure it. For more information, see Amazon Web Services managed policies for Amazon Web Services Outposts and Logging Amazon Web Services Outposts API calls with Amazon Web Services CloudTrail in the Amazon Web Services Outposts User Guide.

Arguments

connection_id: The ID of the connection.

Main.Outposts.get_order — Method

get_order(order_id)
get_order(order_id, params::Dict{String,<:Any})

Gets information about the specified order.

Arguments

order_id: The ID of the order.

Main.Outposts.get_outpost — Method

get_outpost(outpost_id)
get_outpost(outpost_id, params::Dict{String,<:Any})

Gets information about the specified Outpost.

Arguments

outpost_id: The ID or the Amazon Resource Name (ARN) of the Outpost.

Main.Outposts.get_outpost_instance_types — Method

get_outpost_instance_types(outpost_id)
get_outpost_instance_types(outpost_id, params::Dict{String,<:Any})

Gets the instance types for the specified Outpost.

Arguments

outpost_id: The ID or the Amazon Resource Name (ARN) of the Outpost.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults":

"NextToken":

Main.Outposts.get_site — Method

get_site(site_id)
get_site(site_id, params::Dict{String,<:Any})

Gets information about the specified Outpost site.

Arguments

site_id: The ID or the Amazon Resource Name (ARN) of the site.

Main.Outposts.get_site_address — Method

get_site_address(address_type, site_id)
get_site_address(address_type, site_id, params::Dict{String,<:Any})

Gets the site address of the specified site.

Arguments

address_type: The type of the address you request.

site_id: The ID or the Amazon Resource Name (ARN) of the site.

Main.Outposts.list_assets — Method

list_assets(outpost_id)
list_assets(outpost_id, params::Dict{String,<:Any})

Lists the hardware assets for the specified Outpost. Use filters to return specific results. If you specify multiple filters, the results include only the resources that match all of the specified filters. For a filter where you can specify multiple values, the results include items that match any of the values that you specify for the filter.

Arguments

outpost_id: The ID or the Amazon Resource Name (ARN) of the Outpost.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"HostIdFilter": Filters the results by the host ID of a Dedicated Host.

"MaxResults":

"NextToken":

"StatusFilter": Filters the results by state.

Main.Outposts.list_catalog_items — Method

list_catalog_items()
list_catalog_items(params::Dict{String,<:Any})

Lists the items in the catalog. Use filters to return specific results. If you specify multiple filters, the results include only the resources that match all of the specified filters. For a filter where you can specify multiple values, the results include items that match any of the values that you specify for the filter.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"EC2FamilyFilter": Filters the results by EC2 family (for example, M5).

"ItemClassFilter": Filters the results by item class.

"MaxResults":

"NextToken":

"SupportedStorageFilter": Filters the results by storage option.

Main.Outposts.list_orders — Method

list_orders()
list_orders(params::Dict{String,<:Any})

Lists the Outpost orders for your Amazon Web Services account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults":

"NextToken":

"OutpostIdentifierFilter": The ID or the Amazon Resource Name (ARN) of the Outpost.

Main.Outposts.list_outposts — Method

list_outposts()
list_outposts(params::Dict{String,<:Any})

Lists the Outposts for your Amazon Web Services account. Use filters to return specific results. If you specify multiple filters, the results include only the resources that match all of the specified filters. For a filter where you can specify multiple values, the results include items that match any of the values that you specify for the filter.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AvailabilityZoneFilter": Filters the results by Availability Zone (for example, us-east-1a).

"AvailabilityZoneIdFilter": Filters the results by AZ ID (for example, use1-az1).

"LifeCycleStatusFilter": Filters the results by the lifecycle status.

"MaxResults":

"NextToken":

Main.Outposts.list_sites — Method

list_sites()
list_sites(params::Dict{String,<:Any})

Lists the Outpost sites for your Amazon Web Services account. Use filters to return specific results. Use filters to return specific results. If you specify multiple filters, the results include only the resources that match all of the specified filters. For a filter where you can specify multiple values, the results include items that match any of the values that you specify for the filter.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults":

"NextToken":

"OperatingAddressCityFilter": Filters the results by city.

"OperatingAddressCountryCodeFilter": Filters the results by country code.

"OperatingAddressStateOrRegionFilter": Filters the results by state or region.

Main.Outposts.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists the tags for the specified resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

Main.Outposts.start_connection — Method

start_connection(asset_id, client_public_key, device_serial_number, network_interface_device_index)
start_connection(asset_id, client_public_key, device_serial_number, network_interface_device_index, params::Dict{String,<:Any})

Amazon Web Services uses this action to install Outpost servers. Starts the connection required for Outpost server installation. Use CloudTrail to monitor this action or Amazon Web Services managed policy for Amazon Web Services Outposts to secure it. For more information, see Amazon Web Services managed policies for Amazon Web Services Outposts and Logging Amazon Web Services Outposts API calls with Amazon Web Services CloudTrail in the Amazon Web Services Outposts User Guide.

Arguments

asset_id: The ID of the Outpost server.

client_public_key: The public key of the client.

device_serial_number: The serial number of the dongle.

network_interface_device_index: The device index of the network interface on the Outpost server.

Main.Outposts.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Adds tags to the specified resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

tags: The tags to add to the resource.

Main.Outposts.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes tags from the specified resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

tag_keys: The tag keys.

Main.Outposts.update_outpost — Method

update_outpost(outpost_id)
update_outpost(outpost_id, params::Dict{String,<:Any})

Updates an Outpost.

Arguments

outpost_id: The ID or the Amazon Resource Name (ARN) of the Outpost.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description":

"Name":

"SupportedHardwareType": The type of hardware for this Outpost.

Main.Outposts.update_site — Method

update_site(site_id)
update_site(site_id, params::Dict{String,<:Any})

Updates the specified site.

Arguments

site_id: The ID or the Amazon Resource Name (ARN) of the site.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description":

"Name":

"Notes": Notes about a site.

Main.Outposts.update_site_address — Method

update_site_address(address, address_type, site_id)
update_site_address(address, address_type, site_id, params::Dict{String,<:Any})

Updates the address of the specified site. You can't update a site address if there is an order in progress. You must wait for the order to complete or cancel the order. You can update the operating address before you place an order at the site, or after all Outposts that belong to the site have been deactivated.

Arguments

address: The address for the site.

address_type: The type of the address.

site_id: The ID or the Amazon Resource Name (ARN) of the site.

Main.Outposts.update_site_rack_physical_properties — Method

update_site_rack_physical_properties(site_id)
update_site_rack_physical_properties(site_id, params::Dict{String,<:Any})

Update the physical and logistical details for a rack at a site. For more information about hardware requirements for racks, see Network readiness checklist in the Amazon Web Services Outposts User Guide. To update a rack at a site with an order of IN_PROGRESS, you must wait for the order to complete or cancel the order.

Arguments

site_id: The ID or the Amazon Resource Name (ARN) of the site.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"FiberOpticCableType": The type of fiber that you will use to attach the Outpost to your network.

"MaximumSupportedWeightLbs": The maximum rack weight that this site can support. NO_LIMIT is over 2000lbs.

"OpticalStandard": The type of optical standard that you will use to attach the Outpost to your network. This field is dependent on uplink speed, fiber type, and distance to the upstream device. For more information about networking requirements for racks, see Network in the Amazon Web Services Outposts User Guide. OPTIC10GBASESR: 10GBASE-SR OPTIC10GBASEIR: 10GBASE-IR OPTIC10GBASELR: 10GBASE-LR OPTIC40GBASESR: 40GBASE-SR OPTIC40GBASEESR: 40GBASE-ESR OPTIC40GBASEIR4LR4L: 40GBASE-IR (LR4L) OPTIC40GBASELR4: 40GBASE-LR4 OPTIC100GBASESR4: 100GBASE-SR4 OPTIC100GBASECWDM4: 100GBASE-CWDM4 OPTIC100GBASELR4: 100GBASE-LR4 OPTIC100GPSM4MSA: 100G PSM4 MSA OPTIC1000BASELX: 1000Base-LX OPTIC1000BASESX : 1000Base-SX

"PowerConnector": The power connector that Amazon Web Services should plan to provide for connections to the hardware. Note the correlation between PowerPhase and PowerConnector. Single-phase AC feed L6-30P – (common in US); 30A; single phase IEC309 (blue) – P+N+E, 6hr; 32 A; single phase Three-phase AC feed AH530P7W (red) – 3P+N+E, 7hr; 30A; three phase AH532P6W (red) – 3P+N+E, 6hr; 32A; three phase

"PowerDrawKva": The power draw, in kVA, available at the hardware placement position for the rack.

"PowerFeedDrop": Indicates whether the power feed comes above or below the rack.

"PowerPhase": The power option that you can provide for hardware. Single-phase AC feed: 200 V to 277 V, 50 Hz or 60 Hz Three-phase AC feed: 346 V to 480 V, 50 Hz or 60 Hz

"UplinkCount": Racks come with two Outpost network devices. Depending on the supported uplink speed at the site, the Outpost network devices provide a variable number of uplinks. Specify the number of uplinks for each Outpost network device that you intend to use to connect the rack to your network. Note the correlation between UplinkGbps and UplinkCount. 1Gbps - Uplinks available: 1, 2, 4, 6, 8 10Gbps - Uplinks available: 1, 2, 4, 8, 12, 16 40 and 100 Gbps- Uplinks available: 1, 2, 4

"UplinkGbps": The uplink speed the rack should support for the connection to the Region.