Panorama

This page documents function available when using the Panorama module, created with @service Panorama.

Index

Main.Panorama.create_application_instance

Main.Panorama.create_job_for_devices

Main.Panorama.create_node_from_template_job

Main.Panorama.create_package

Main.Panorama.create_package_import_job

Main.Panorama.delete_device

Main.Panorama.delete_package

Main.Panorama.deregister_package_version

Main.Panorama.describe_application_instance

Main.Panorama.describe_application_instance_details

Main.Panorama.describe_device

Main.Panorama.describe_device_job

Main.Panorama.describe_node

Main.Panorama.describe_node_from_template_job

Main.Panorama.describe_package

Main.Panorama.describe_package_import_job

Main.Panorama.describe_package_version

Main.Panorama.list_application_instance_dependencies

Main.Panorama.list_application_instance_node_instances

Main.Panorama.list_application_instances

Main.Panorama.list_devices

Main.Panorama.list_devices_jobs

Main.Panorama.list_node_from_template_jobs

Main.Panorama.list_nodes

Main.Panorama.list_package_import_jobs

Main.Panorama.list_packages

Main.Panorama.list_tags_for_resource

Main.Panorama.provision_device

Main.Panorama.register_package_version

Main.Panorama.remove_application_instance

Main.Panorama.tag_resource

Main.Panorama.untag_resource

Main.Panorama.update_device_metadata

Documentation

Main.Panorama.create_application_instance — Method

create_application_instance(default_runtime_context_device, manifest_payload)
create_application_instance(default_runtime_context_device, manifest_payload, params::Dict{String,<:Any})

Creates an application instance and deploys it to a device.

Arguments

default_runtime_context_device: A device's ID.

manifest_payload: The application's manifest document.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ApplicationInstanceIdToReplace": The ID of an application instance to replace with the new instance.

"Description": A description for the application instance.

"ManifestOverridesPayload": Setting overrides for the application manifest.

"Name": A name for the application instance.

"RuntimeRoleArn": The ARN of a runtime role for the application instance.

"Tags": Tags for the application instance.

Main.Panorama.create_job_for_devices — Method

create_job_for_devices(device_ids, device_job_config, job_type)
create_job_for_devices(device_ids, device_job_config, job_type, params::Dict{String,<:Any})

Creates a job to run on one or more devices.

Arguments

device_ids: IDs of target devices.

device_job_config: Configuration settings for the job.

job_type: The type of job to run.

Main.Panorama.create_node_from_template_job — Method

create_node_from_template_job(node_name, output_package_name, output_package_version, template_parameters, template_type)
create_node_from_template_job(node_name, output_package_name, output_package_version, template_parameters, template_type, params::Dict{String,<:Any})

Creates a camera stream node.

Arguments

node_name: A name for the node.

output_package_name: An output package name for the node.

output_package_version: An output package version for the node.

template_parameters: Template parameters for the node.

template_type: The type of node.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"JobTags": Tags for the job.

"NodeDescription": A description for the node.

Main.Panorama.create_package — Method

create_package(package_name)
create_package(package_name, params::Dict{String,<:Any})

Creates a package and storage location in an Amazon S3 access point.

Arguments

package_name: A name for the package.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Tags": Tags for the package.

Main.Panorama.create_package_import_job — Method

create_package_import_job(client_token, input_config, job_type, output_config)
create_package_import_job(client_token, input_config, job_type, output_config, params::Dict{String,<:Any})

Imports a node package.

Arguments

client_token: A client token for the package import job.

input_config: An input config for the package import job.

job_type: A job type for the package import job.

output_config: An output config for the package import job.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"JobTags": Tags for the package import job.

Main.Panorama.delete_device — Method

delete_device(device_id)
delete_device(device_id, params::Dict{String,<:Any})

Deletes a device.

Arguments

device_id: The device's ID.

Main.Panorama.delete_package — Method

delete_package(package_id)
delete_package(package_id, params::Dict{String,<:Any})

Deletes a package. To delete a package, you need permission to call s3:DeleteObject in addition to permissions for the AWS Panorama API.

Arguments

package_id: The package's ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ForceDelete": Delete the package even if it has artifacts stored in its access point. Deletes the package's artifacts from Amazon S3.

Main.Panorama.deregister_package_version — Method

deregister_package_version(package_id, package_version, patch_version)
deregister_package_version(package_id, package_version, patch_version, params::Dict{String,<:Any})

Deregisters a package version.

Arguments

package_id: A package ID.

package_version: A package version.

patch_version: A patch version.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"OwnerAccount": An owner account.

"UpdatedLatestPatchVersion": If the version was marked latest, the new version to maker as latest.

Main.Panorama.describe_application_instance — Method

describe_application_instance(application_instance_id)
describe_application_instance(application_instance_id, params::Dict{String,<:Any})

Returns information about an application instance on a device.

Arguments

application_instance_id: The application instance's ID.

Main.Panorama.describe_application_instance_details — Method

describe_application_instance_details(application_instance_id)
describe_application_instance_details(application_instance_id, params::Dict{String,<:Any})

Returns information about an application instance's configuration manifest.

Arguments

application_instance_id: The application instance's ID.

Main.Panorama.describe_device — Method

describe_device(device_id)
describe_device(device_id, params::Dict{String,<:Any})

Returns information about a device.

Arguments

device_id: The device's ID.

Main.Panorama.describe_device_job — Method

describe_device_job(job_id)
describe_device_job(job_id, params::Dict{String,<:Any})

Returns information about a device job.

Arguments

job_id: The job's ID.

Main.Panorama.describe_node — Method

describe_node(node_id)
describe_node(node_id, params::Dict{String,<:Any})

Returns information about a node.

Arguments

node_id: The node's ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"OwnerAccount": The account ID of the node's owner.

Main.Panorama.describe_node_from_template_job — Method

describe_node_from_template_job(job_id)
describe_node_from_template_job(job_id, params::Dict{String,<:Any})

Returns information about a job to create a camera stream node.

Arguments

job_id: The job's ID.

Main.Panorama.describe_package — Method

describe_package(package_id)
describe_package(package_id, params::Dict{String,<:Any})

Returns information about a package.

Arguments

package_id: The package's ID.

Main.Panorama.describe_package_import_job — Method

describe_package_import_job(job_id)
describe_package_import_job(job_id, params::Dict{String,<:Any})

Returns information about a package import job.

Arguments

job_id: The job's ID.

Main.Panorama.describe_package_version — Method

describe_package_version(package_id, package_version)
describe_package_version(package_id, package_version, params::Dict{String,<:Any})

Returns information about a package version.

Arguments

package_id: The version's ID.

package_version: The version's version.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"OwnerAccount": The version's owner account.

"PatchVersion": The version's patch version.

Main.Panorama.list_application_instance_dependencies — Method

list_application_instance_dependencies(application_instance_id)
list_application_instance_dependencies(application_instance_id, params::Dict{String,<:Any})

Returns a list of application instance dependencies.

Arguments

application_instance_id: The application instance's ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of application instance dependencies to return in one page of results.

"nextToken": Specify the pagination token from a previous request to retrieve the next page of results.

Main.Panorama.list_application_instance_node_instances — Method

list_application_instance_node_instances(application_instance_id)
list_application_instance_node_instances(application_instance_id, params::Dict{String,<:Any})

Returns a list of application node instances.

Arguments

application_instance_id: The node instances' application instance ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of node instances to return in one page of results.

"nextToken": Specify the pagination token from a previous request to retrieve the next page of results.

Main.Panorama.list_application_instances — Method

list_application_instances()
list_application_instances(params::Dict{String,<:Any})

Returns a list of application instances.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"deviceId": The application instances' device ID.

"maxResults": The maximum number of application instances to return in one page of results.

"nextToken": Specify the pagination token from a previous request to retrieve the next page of results.

"statusFilter": Only include instances with a specific status.

Main.Panorama.list_devices — Method

list_devices()
list_devices(params::Dict{String,<:Any})

Returns a list of devices.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DeviceAggregatedStatusFilter": Filter based on a device's status.

"MaxResults": The maximum number of devices to return in one page of results.

"NameFilter": Filter based on device's name. Prefixes supported.

"NextToken": Specify the pagination token from a previous request to retrieve the next page of results.

"SortBy": The target column to be sorted on. Default column sort is CREATED_TIME.

"SortOrder": The sorting order for the returned list. SortOrder is DESCENDING by default based on CREATED_TIME. Otherwise, SortOrder is ASCENDING.

Main.Panorama.list_devices_jobs — Method

list_devices_jobs()
list_devices_jobs(params::Dict{String,<:Any})

Returns a list of jobs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DeviceId": Filter results by the job's target device ID.

"MaxResults": The maximum number of device jobs to return in one page of results.

"NextToken": Specify the pagination token from a previous request to retrieve the next page of results.

Main.Panorama.list_node_from_template_jobs — Method

list_node_from_template_jobs()
list_node_from_template_jobs(params::Dict{String,<:Any})

Returns a list of camera stream node jobs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of node from template jobs to return in one page of results.

"NextToken": Specify the pagination token from a previous request to retrieve the next page of results.

Main.Panorama.list_nodes — Method

list_nodes()
list_nodes(params::Dict{String,<:Any})

Returns a list of nodes.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"category": Search for nodes by category.

"maxResults": The maximum number of nodes to return in one page of results.

"nextToken": Specify the pagination token from a previous request to retrieve the next page of results.

"ownerAccount": Search for nodes by the account ID of the nodes' owner.

"packageName": Search for nodes by name.

"packageVersion": Search for nodes by version.

"patchVersion": Search for nodes by patch version.

Main.Panorama.list_package_import_jobs — Method

list_package_import_jobs()
list_package_import_jobs(params::Dict{String,<:Any})

Returns a list of package import jobs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of package import jobs to return in one page of results.

"NextToken": Specify the pagination token from a previous request to retrieve the next page of results.

Main.Panorama.list_packages — Method

list_packages()
list_packages(params::Dict{String,<:Any})

Returns a list of packages.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of packages to return in one page of results.

"nextToken": Specify the pagination token from a previous request to retrieve the next page of results.

Main.Panorama.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Returns a list of tags for a resource.

Arguments

resource_arn: The resource's ARN.

Main.Panorama.provision_device — Method

provision_device(name)
provision_device(name, params::Dict{String,<:Any})

Creates a device and returns a configuration archive. The configuration archive is a ZIP file that contains a provisioning certificate that is valid for 5 minutes. Name the configuration archive certificates-omni_device-name.zip and transfer it to the device within 5 minutes. Use the included USB storage device and connect it to the USB 3.0 port next to the HDMI output.

Arguments

name: A name for the device.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A description for the device.

"NetworkingConfiguration": A networking configuration for the device.

"Tags": Tags for the device.

Main.Panorama.register_package_version — Method

register_package_version(package_id, package_version, patch_version)
register_package_version(package_id, package_version, patch_version, params::Dict{String,<:Any})

Registers a package version.

Arguments

package_id: A package ID.

package_version: A package version.

patch_version: A patch version.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MarkLatest": Whether to mark the new version as the latest version.

"OwnerAccount": An owner account.

Main.Panorama.remove_application_instance — Method

remove_application_instance(application_instance_id)
remove_application_instance(application_instance_id, params::Dict{String,<:Any})

Removes an application instance.

Arguments

application_instance_id: An application instance ID.

Main.Panorama.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Tags a resource.

Arguments

resource_arn: The resource's ARN.

tags: Tags for the resource.

Main.Panorama.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes tags from a resource.

Arguments

resource_arn: The resource's ARN.

tag_keys: Tag keys to remove.

Main.Panorama.update_device_metadata — Method

update_device_metadata(device_id)
update_device_metadata(device_id, params::Dict{String,<:Any})

Updates a device's metadata.

Arguments

device_id: The device's ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A description for the device.