Pcs

This page documents function available when using the Pcs module, created with @service Pcs.

Index

Main.Pcs.create_cluster

Main.Pcs.create_compute_node_group

Main.Pcs.create_queue

Main.Pcs.delete_cluster

Main.Pcs.delete_compute_node_group

Main.Pcs.delete_queue

Main.Pcs.get_cluster

Main.Pcs.get_compute_node_group

Main.Pcs.get_queue

Main.Pcs.list_clusters

Main.Pcs.list_compute_node_groups

Main.Pcs.list_queues

Main.Pcs.list_tags_for_resource

Main.Pcs.register_compute_node_group_instance

Main.Pcs.tag_resource

Main.Pcs.untag_resource

Main.Pcs.update_compute_node_group

Main.Pcs.update_queue

Documentation

Main.Pcs.create_cluster — Function

create_cluster(cluster_name, networking, scheduler, size)
create_cluster(cluster_name, networking, scheduler, size, params::Dict{String,<:Any})

Creates a cluster in your account. Amazon Web Services PCS creates the cluster controller in a service-owned account. The cluster controller communicates with the cluster resources in your account. The subnets and security groups for the cluster must already exist before you use this API action. It takes time for Amazon Web Services PCS to create the cluster. The cluster is in a Creating state until it is ready to use. There can only be 1 cluster in a Creating state per Amazon Web Services Region per Amazon Web Services account. CreateCluster fails with a ServiceQuotaExceededException if there is already a cluster in a Creating state.

Arguments

cluster_name: A name to identify the cluster. Example: MyCluster

networking: The networking configuration used to set up the cluster's control plane.

scheduler: The cluster management and job scheduling software associated with the cluster.

size: A value that determines the maximum number of compute nodes in the cluster and the maximum number of jobs (active and queued). SMALL: 32 compute nodes and 256 jobs MEDIUM: 512 compute nodes and 8192 jobs LARGE: 2048 compute nodes and 16,384 jobs

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, the subsequent retries with the same client token return the result from the original successful request and they have no additional effect. If you don't specify a client token, the CLI and SDK automatically generate 1 for you.

"slurmConfiguration": Additional options related to the Slurm scheduler.

"tags": 1 or more tags added to the resource. Each tag consists of a tag key and tag value. The tag value is optional and can be an empty string.

Main.Pcs.create_compute_node_group — Function

create_compute_node_group(cluster_identifier, compute_node_group_name, custom_launch_template, iam_instance_profile_arn, instance_configs, scaling_configuration, subnet_ids)
create_compute_node_group(cluster_identifier, compute_node_group_name, custom_launch_template, iam_instance_profile_arn, instance_configs, scaling_configuration, subnet_ids, params::Dict{String,<:Any})

Creates a managed set of compute nodes. You associate a compute node group with a cluster through 1 or more Amazon Web Services PCS queues or as part of the login fleet. A compute node group includes the definition of the compute properties and lifecycle management. Amazon Web Services PCS uses the information you provide to this API action to launch compute nodes in your account. You can only specify subnets in the same Amazon VPC as your cluster. You receive billing charges for the compute nodes that Amazon Web Services PCS launches in your account. You must already have a launch template before you call this API. For more information, see Launch an instance from a launch template in the Amazon Elastic Compute Cloud User Guide for Linux Instances.

Arguments

cluster_identifier: The name or ID of the cluster to create a compute node group in.

compute_node_group_name: A name to identify the cluster. Example: MyCluster

custom_launch_template:

iam_instance_profile_arn: The Amazon Resource Name (ARN) of the IAM instance profile used to pass an IAM role when launching EC2 instances. The role contained in your instance profile must have pcs:RegisterComputeNodeGroupInstance permissions attached in order to provision instances correctly. The resource identifier of the ARN must start with AWSPCS. For example, arn:aws:iam:123456789012:instance-profile/AWSPCSMyComputeNodeInstanceProfile.

instance_configs: A list of EC2 instance configurations that Amazon Web Services PCS can provision in the compute node group.

scaling_configuration: Specifies the boundaries of the compute node group auto scaling.

subnet_ids: The list of subnet IDs where the compute node group launches instances. Subnets must be in the same VPC as the cluster.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"amiId": The ID of the Amazon Machine Image (AMI) that Amazon Web Services PCS uses to launch compute nodes (Amazon EC2 instances). If you don't provide this value, Amazon Web Services PCS uses the AMI ID specified in the custom launch template.

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, the subsequent retries with the same client token return the result from the original successful request and they have no additional effect. If you don't specify a client token, the CLI and SDK automatically generate 1 for you.

"purchaseOption": Specifies how EC2 instances are purchased on your behalf. Amazon Web Services PCS supports On-Demand and Spot instances. For more information, see Instance purchasing options in the Amazon Elastic Compute Cloud User Guide. If you don't provide this option, it defaults to On-Demand.

"slurmConfiguration": Additional options related to the Slurm scheduler.

"spotOptions":

"tags": 1 or more tags added to the resource. Each tag consists of a tag key and tag value. The tag value is optional and can be an empty string.

Main.Pcs.create_queue — Function

create_queue(cluster_identifier, queue_name)
create_queue(cluster_identifier, queue_name, params::Dict{String,<:Any})

Creates a job queue. You must associate 1 or more compute node groups with the queue. You can associate 1 compute node group with multiple queues.

Arguments

cluster_identifier: The name or ID of the cluster for which to create a queue.

queue_name: A name to identify the queue.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, the subsequent retries with the same client token return the result from the original successful request and they have no additional effect. If you don't specify a client token, the CLI and SDK automatically generate 1 for you.

"computeNodeGroupConfigurations": The list of compute node group configurations to associate with the queue. Queues assign jobs to associated compute node groups.

"tags": 1 or more tags added to the resource. Each tag consists of a tag key and tag value. The tag value is optional and can be an empty string.

Main.Pcs.delete_cluster — Function

delete_cluster(cluster_identifier)
delete_cluster(cluster_identifier, params::Dict{String,<:Any})

Deletes a cluster and all its linked resources. You must delete all queues and compute node groups associated with the cluster before you can delete the cluster.

Arguments

cluster_identifier: The name or ID of the cluster to delete.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, the subsequent retries with the same client token return the result from the original successful request and they have no additional effect. If you don't specify a client token, the CLI and SDK automatically generate 1 for you.

Main.Pcs.delete_compute_node_group — Function

delete_compute_node_group(cluster_identifier, compute_node_group_identifier)
delete_compute_node_group(cluster_identifier, compute_node_group_identifier, params::Dict{String,<:Any})

Deletes a compute node group. You must delete all queues associated with the compute node group first.

Arguments

cluster_identifier: The name or ID of the cluster of the compute node group.

compute_node_group_identifier: The name or ID of the compute node group to delete.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, the subsequent retries with the same client token return the result from the original successful request and they have no additional effect. If you don't specify a client token, the CLI and SDK automatically generate 1 for you.

Main.Pcs.delete_queue — Function

delete_queue(cluster_identifier, queue_identifier)
delete_queue(cluster_identifier, queue_identifier, params::Dict{String,<:Any})

Deletes a job queue. If the compute node group associated with this queue isn't associated with any other queues, Amazon Web Services PCS terminates all the compute nodes for this queue.

Arguments

cluster_identifier: The name or ID of the cluster of the queue.

queue_identifier: The name or ID of the queue to delete.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, the subsequent retries with the same client token return the result from the original successful request and they have no additional effect. If you don't specify a client token, the CLI and SDK automatically generate 1 for you.

Main.Pcs.get_cluster — Function

get_cluster(cluster_identifier)
get_cluster(cluster_identifier, params::Dict{String,<:Any})

Returns detailed information about a running cluster in your account. This API action provides networking information, endpoint information for communication with the scheduler, and provisioning status.

Arguments

cluster_identifier: The name or ID of the cluster of the queue.

Main.Pcs.get_compute_node_group — Function

get_compute_node_group(cluster_identifier, compute_node_group_identifier)
get_compute_node_group(cluster_identifier, compute_node_group_identifier, params::Dict{String,<:Any})

Returns detailed information about a compute node group. This API action provides networking information, EC2 instance type, compute node group status, and scheduler (such as Slurm) configuration.

Arguments

cluster_identifier: The name or ID of the cluster.

compute_node_group_identifier: The name or ID of the compute node group.

Main.Pcs.get_queue — Function

get_queue(cluster_identifier, queue_identifier)
get_queue(cluster_identifier, queue_identifier, params::Dict{String,<:Any})

Returns detailed information about a queue. The information includes the compute node groups that the queue uses to schedule jobs.

Arguments

cluster_identifier: The name or ID of the cluster of the queue.

queue_identifier: The name or ID of the queue.

Main.Pcs.list_clusters — Function

list_clusters()
list_clusters(params::Dict{String,<:Any})

Returns a list of running clusters in your account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results that are returned per call. You can use nextToken to obtain further pages of results. The default is 10 results, and the maximum allowed page size is 100 results. A value of 0 uses the default.

"nextToken": The value of nextToken is a unique pagination token for each page of results returned. If nextToken is returned, there are more results available. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token returns an HTTP 400 InvalidToken error.

Main.Pcs.list_compute_node_groups — Function

list_compute_node_groups(cluster_identifier)
list_compute_node_groups(cluster_identifier, params::Dict{String,<:Any})

Returns a list of all compute node groups associated with a cluster.

Arguments

cluster_identifier: The name or ID of the cluster to list compute node groups for.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results that are returned per call. You can use nextToken to obtain further pages of results. The default is 10 results, and the maximum allowed page size is 100 results. A value of 0 uses the default.

"nextToken": The value of nextToken is a unique pagination token for each page of results returned. If nextToken is returned, there are more results available. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token returns an HTTP 400 InvalidToken error.

Main.Pcs.list_queues — Function

list_queues(cluster_identifier)
list_queues(cluster_identifier, params::Dict{String,<:Any})

Returns a list of all queues associated with a cluster.

Arguments

cluster_identifier: The name or ID of the cluster to list queues for.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results that are returned per call. You can use nextToken to obtain further pages of results. The default is 10 results, and the maximum allowed page size is 100 results. A value of 0 uses the default.

"nextToken": The value of nextToken is a unique pagination token for each page of results returned. If nextToken is returned, there are more results available. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token returns an HTTP 400 InvalidToken error.

Main.Pcs.list_tags_for_resource — Function

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Returns a list of all tags on an Amazon Web Services PCS resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource for which to list tags.

Main.Pcs.register_compute_node_group_instance — Function

register_compute_node_group_instance(bootstrap_id, cluster_identifier)
register_compute_node_group_instance(bootstrap_id, cluster_identifier, params::Dict{String,<:Any})

This API action isn't intended for you to use. Amazon Web Services PCS uses this API action to register the compute nodes it launches in your account.

Arguments

bootstrap_id: The client-generated token to allow for retries.

cluster_identifier: The name or ID of the cluster to register the compute node group instance in.

Main.Pcs.tag_resource — Function

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Adds or edits tags on an Amazon Web Services PCS resource. Each tag consists of a tag key and a tag value. The tag key and tag value are case-sensitive strings. The tag value can be an empty (null) string. To add a tag, specify a new tag key and a tag value. To edit a tag, specify an existing tag key and a new tag value.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

tags: 1 or more tags added to the resource. Each tag consists of a tag key and tag value. The tag value is optional and can be an empty string.

Main.Pcs.untag_resource — Function

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Deletes tags from an Amazon Web Services PCS resource. To delete a tag, specify the tag key and the Amazon Resource Name (ARN) of the Amazon Web Services PCS resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

tag_keys: 1 or more tag keys to remove from the resource. Specify only tag keys and not tag values.

Main.Pcs.update_compute_node_group — Function

update_compute_node_group(cluster_identifier, compute_node_group_identifier)
update_compute_node_group(cluster_identifier, compute_node_group_identifier, params::Dict{String,<:Any})

Updates a compute node group. You can update many of the fields related to your compute node group including the configurations for networking, compute nodes, and settings specific to your scheduler (such as Slurm).

Arguments

cluster_identifier: The name or ID of the cluster of the compute node group.

compute_node_group_identifier: The name or ID of the compute node group.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"amiId": The ID of the Amazon Machine Image (AMI) that Amazon Web Services PCS uses to launch instances. If not provided, Amazon Web Services PCS uses the AMI ID specified in the custom launch template.

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, the subsequent retries with the same client token return the result from the original successful request and they have no additional effect. If you don't specify a client token, the CLI and SDK automatically generate 1 for you.

"customLaunchTemplate":

"iamInstanceProfileArn": The Amazon Resource Name (ARN) of the IAM instance profile used to pass an IAM role when launching EC2 instances. The role contained in your instance profile must have pcs:RegisterComputeNodeGroupInstance permissions attached to provision instances correctly.

"purchaseOption": Specifies how EC2 instances are purchased on your behalf. Amazon Web Services PCS supports On-Demand and Spot instances. For more information, see Instance purchasing options in the Amazon Elastic Compute Cloud User Guide. If you don't provide this option, it defaults to On-Demand.

"scalingConfiguration": Specifies the boundaries of the compute node group auto scaling.

"slurmConfiguration": Additional options related to the Slurm scheduler.

"spotOptions":

"subnetIds": The list of subnet IDs where the compute node group provisions instances. The subnets must be in the same VPC as the cluster.

Main.Pcs.update_queue — Function

update_queue(cluster_identifier, queue_identifier)
update_queue(cluster_identifier, queue_identifier, params::Dict{String,<:Any})

Updates the compute node group configuration of a queue. Use this API to change the compute node groups that the queue can send jobs to.

Arguments

cluster_identifier: The name or ID of the cluster of the queue.

queue_identifier: The name or ID of the queue.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, the subsequent retries with the same client token return the result from the original successful request and they have no additional effect. If you don't specify a client token, the CLI and SDK automatically generate 1 for you.

"computeNodeGroupConfigurations": The list of compute node group configurations to associate with the queue. Queues assign jobs to associated compute node groups.