Pinpoint Sms Voice V2

This page documents function available when using the Pinpoint_Sms_Voice_V2 module, created with @service Pinpoint_Sms_Voice_V2.

Index

Main.Pinpoint_Sms_Voice_V2.associate_origination_identity

Main.Pinpoint_Sms_Voice_V2.create_configuration_set

Main.Pinpoint_Sms_Voice_V2.create_event_destination

Main.Pinpoint_Sms_Voice_V2.create_opt_out_list

Main.Pinpoint_Sms_Voice_V2.create_pool

Main.Pinpoint_Sms_Voice_V2.delete_configuration_set

Main.Pinpoint_Sms_Voice_V2.delete_default_message_type

Main.Pinpoint_Sms_Voice_V2.delete_default_sender_id

Main.Pinpoint_Sms_Voice_V2.delete_event_destination

Main.Pinpoint_Sms_Voice_V2.delete_keyword

Main.Pinpoint_Sms_Voice_V2.delete_opt_out_list

Main.Pinpoint_Sms_Voice_V2.delete_opted_out_number

Main.Pinpoint_Sms_Voice_V2.delete_pool

Main.Pinpoint_Sms_Voice_V2.delete_text_message_spend_limit_override

Main.Pinpoint_Sms_Voice_V2.delete_voice_message_spend_limit_override

Main.Pinpoint_Sms_Voice_V2.describe_account_attributes

Main.Pinpoint_Sms_Voice_V2.describe_account_limits

Main.Pinpoint_Sms_Voice_V2.describe_configuration_sets

Main.Pinpoint_Sms_Voice_V2.describe_keywords

Main.Pinpoint_Sms_Voice_V2.describe_opt_out_lists

Main.Pinpoint_Sms_Voice_V2.describe_opted_out_numbers

Main.Pinpoint_Sms_Voice_V2.describe_phone_numbers

Main.Pinpoint_Sms_Voice_V2.describe_pools

Main.Pinpoint_Sms_Voice_V2.describe_sender_ids

Main.Pinpoint_Sms_Voice_V2.describe_spend_limits

Main.Pinpoint_Sms_Voice_V2.disassociate_origination_identity

Main.Pinpoint_Sms_Voice_V2.list_pool_origination_identities

Main.Pinpoint_Sms_Voice_V2.list_tags_for_resource

Main.Pinpoint_Sms_Voice_V2.put_keyword

Main.Pinpoint_Sms_Voice_V2.put_opted_out_number

Main.Pinpoint_Sms_Voice_V2.release_phone_number

Main.Pinpoint_Sms_Voice_V2.request_phone_number

Main.Pinpoint_Sms_Voice_V2.send_text_message

Main.Pinpoint_Sms_Voice_V2.send_voice_message

Main.Pinpoint_Sms_Voice_V2.set_default_message_type

Main.Pinpoint_Sms_Voice_V2.set_default_sender_id

Main.Pinpoint_Sms_Voice_V2.set_text_message_spend_limit_override

Main.Pinpoint_Sms_Voice_V2.set_voice_message_spend_limit_override

Main.Pinpoint_Sms_Voice_V2.tag_resource

Main.Pinpoint_Sms_Voice_V2.untag_resource

Main.Pinpoint_Sms_Voice_V2.update_event_destination

Main.Pinpoint_Sms_Voice_V2.update_phone_number

Main.Pinpoint_Sms_Voice_V2.update_pool

Documentation

Main.Pinpoint_Sms_Voice_V2.associate_origination_identity — Method

associate_origination_identity(iso_country_code, origination_identity, pool_id)
associate_origination_identity(iso_country_code, origination_identity, pool_id, params::Dict{String,<:Any})

Associates the specified origination identity with a pool. If the origination identity is a phone number and is already associated with another pool, an Error is returned. A sender ID can be associated with multiple pools. If the origination identity configuration doesn't match the pool's configuration, an Error is returned.

Arguments

iso_country_code: The new two-character code, in ISO 3166-1 alpha-2 format, for the country or region of the origination identity.

origination_identity: The origination identity to use, such as PhoneNumberId, PhoneNumberArn, SenderId, or SenderIdArn. You can use DescribePhoneNumbers to find the values for PhoneNumberId and PhoneNumberArn, while DescribeSenderIds can be used to get the values for SenderId and SenderIdArn.

pool_id: The pool to update with the new Identity. This value can be either the PoolId or PoolArn, and you can find these values using DescribePools.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If you don't specify a client token, a randomly generated token is used for the request to ensure idempotency.

Main.Pinpoint_Sms_Voice_V2.create_configuration_set — Method

create_configuration_set(configuration_set_name)
create_configuration_set(configuration_set_name, params::Dict{String,<:Any})

Creates a new configuration set. After you create the configuration set, you can add one or more event destinations to it. A configuration set is a set of rules that you apply to the SMS and voice messages that you send. When you send a message, you can optionally specify a single configuration set.

Arguments

configuration_set_name: The name to use for the new configuration set.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If you don't specify a client token, a randomly generated token is used for the request to ensure idempotency.

"Tags": An array of key and value pair tags that's associated with the new configuration set.

Main.Pinpoint_Sms_Voice_V2.create_event_destination — Method

create_event_destination(configuration_set_name, event_destination_name, matching_event_types)
create_event_destination(configuration_set_name, event_destination_name, matching_event_types, params::Dict{String,<:Any})

Creates a new event destination in a configuration set. An event destination is a location where you send message events. The event options are Amazon CloudWatch, Amazon Kinesis Data Firehose, or Amazon SNS. For example, when a message is delivered successfully, you can send information about that event to an event destination, or send notifications to endpoints that are subscribed to an Amazon SNS topic. Each configuration set can contain between 0 and 5 event destinations. Each event destination can contain a reference to a single destination, such as a CloudWatch or Kinesis Data Firehose destination.

Arguments

configuration_set_name: Either the name of the configuration set or the configuration set ARN to apply event logging to. The ConfigurateSetName and ConfigurationSetArn can be found using the DescribeConfigurationSets action.

event_destination_name: The name that identifies the event destination.

matching_event_types: An array of event types that determine which events to log. If "ALL" is used, then Amazon Pinpoint logs every event type.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If you don't specify a client token, a randomly generated token is used for the request to ensure idempotency.

"CloudWatchLogsDestination": An object that contains information about an event destination for logging to Amazon CloudWatch logs.

"KinesisFirehoseDestination": An object that contains information about an event destination for logging to Amazon Kinesis Data Firehose.

"SnsDestination": An object that contains information about an event destination for logging to Amazon SNS.

Main.Pinpoint_Sms_Voice_V2.create_opt_out_list — Method

create_opt_out_list(opt_out_list_name)
create_opt_out_list(opt_out_list_name, params::Dict{String,<:Any})

Creates a new opt-out list. If the opt-out list name already exists, an Error is returned. An opt-out list is a list of phone numbers that are opted out, meaning you can't send SMS or voice messages to them. If end user replies with the keyword "STOP," an entry for the phone number is added to the opt-out list. In addition to STOP, your recipients can use any supported opt-out keyword, such as CANCEL or OPTOUT. For a list of supported opt-out keywords, see SMS opt out in the Amazon Pinpoint User Guide.

Arguments

opt_out_list_name: The name of the new OptOutList.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If you don't specify a client token, a randomly generated token is used for the request to ensure idempotency.

"Tags": An array of tags (key and value pairs) to associate with the new OptOutList.

Main.Pinpoint_Sms_Voice_V2.create_pool — Method

create_pool(iso_country_code, message_type, origination_identity)
create_pool(iso_country_code, message_type, origination_identity, params::Dict{String,<:Any})

Creates a new pool and associates the specified origination identity to the pool. A pool can include one or more phone numbers and SenderIds that are associated with your Amazon Web Services account. The new pool inherits its configuration from the specified origination identity. This includes keywords, message type, opt-out list, two-way configuration, and self-managed opt-out configuration. Deletion protection isn't inherited from the origination identity and defaults to false. If the origination identity is a phone number and is already associated with another pool, an Error is returned. A sender ID can be associated with multiple pools.

Arguments

iso_country_code: The new two-character code, in ISO 3166-1 alpha-2 format, for the country or region of the new pool.

message_type: The type of message. Valid values are TRANSACTIONAL for messages that are critical or time-sensitive and PROMOTIONAL for messages that aren't critical or time-sensitive.

origination_identity: The origination identity to use such as a PhoneNumberId, PhoneNumberArn, SenderId or SenderIdArn. You can use DescribePhoneNumbers to find the values for PhoneNumberId and PhoneNumberArn while DescribeSenderIds can be used to get the values for SenderId and SenderIdArn.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If you don't specify a client token, a randomly generated token is used for the request to ensure idempotency.

"DeletionProtectionEnabled": By default this is set to false. When set to true the pool can't be deleted. You can change this value using the UpdatePool action.

"Tags": An array of tags (key and value pairs) associated with the pool.

Main.Pinpoint_Sms_Voice_V2.delete_configuration_set — Method

delete_configuration_set(configuration_set_name)
delete_configuration_set(configuration_set_name, params::Dict{String,<:Any})

Deletes an existing configuration set. A configuration set is a set of rules that you apply to voice and SMS messages that you send. In a configuration set, you can specify a destination for specific types of events related to voice and SMS messages.

Arguments

configuration_set_name: The name of the configuration set or the configuration set ARN that you want to delete. The ConfigurationSetName and ConfigurationSetArn can be found using the DescribeConfigurationSets action.

Main.Pinpoint_Sms_Voice_V2.delete_default_message_type — Method

delete_default_message_type(configuration_set_name)
delete_default_message_type(configuration_set_name, params::Dict{String,<:Any})

Deletes an existing default message type on a configuration set. A message type is a type of messages that you plan to send. If you send account-related messages or time-sensitive messages such as one-time passcodes, choose Transactional. If you plan to send messages that contain marketing material or other promotional content, choose Promotional. This setting applies to your entire Amazon Web Services account.

Arguments

configuration_set_name: The name of the configuration set or the configuration set Amazon Resource Name (ARN) to delete the default message type from. The ConfigurationSetName and ConfigurationSetArn can be found using the DescribeConfigurationSets action.

Main.Pinpoint_Sms_Voice_V2.delete_default_sender_id — Method

delete_default_sender_id(configuration_set_name)
delete_default_sender_id(configuration_set_name, params::Dict{String,<:Any})

Deletes an existing default sender ID on a configuration set. A default sender ID is the identity that appears on recipients' devices when they receive SMS messages. Support for sender ID capabilities varies by country or region.

Arguments

configuration_set_name: The name of the configuration set or the configuration set Amazon Resource Name (ARN) to delete the default sender ID from. The ConfigurationSetName and ConfigurationSetArn can be found using the DescribeConfigurationSets action.

Main.Pinpoint_Sms_Voice_V2.delete_event_destination — Method

delete_event_destination(configuration_set_name, event_destination_name)
delete_event_destination(configuration_set_name, event_destination_name, params::Dict{String,<:Any})

Deletes an existing event destination. An event destination is a location where you send response information about the messages that you send. For example, when a message is delivered successfully, you can send information about that event to an Amazon CloudWatch destination, or send notifications to endpoints that are subscribed to an Amazon SNS topic.

Arguments

configuration_set_name: The name of the configuration set or the configuration set's Amazon Resource Name (ARN) to remove the event destination from. The ConfigurateSetName and ConfigurationSetArn can be found using the DescribeConfigurationSets action.

event_destination_name: The name of the event destination to delete.

Main.Pinpoint_Sms_Voice_V2.delete_keyword — Method

delete_keyword(keyword, origination_identity)
delete_keyword(keyword, origination_identity, params::Dict{String,<:Any})

Deletes an existing keyword from an origination phone number or pool. A keyword is a word that you can search for on a particular phone number or pool. It is also a specific word or phrase that an end user can send to your number to elicit a response, such as an informational message or a special offer. When your number receives a message that begins with a keyword, Amazon Pinpoint responds with a customizable message. Keywords "HELP" and "STOP" can't be deleted or modified.

Arguments

keyword: The keyword to delete.

origination_identity: The origination identity to use such as a PhoneNumberId, PhoneNumberArn, PoolId or PoolArn. You can use DescribePhoneNumbers to find the values for PhoneNumberId and PhoneNumberArn and DescribePools to find the values of PoolId and PoolArn.

Main.Pinpoint_Sms_Voice_V2.delete_opt_out_list — Method

delete_opt_out_list(opt_out_list_name)
delete_opt_out_list(opt_out_list_name, params::Dict{String,<:Any})

Deletes an existing opt-out list. All opted out phone numbers in the opt-out list are deleted. If the specified opt-out list name doesn't exist or is in-use by an origination phone number or pool, an Error is returned.

Arguments

opt_out_list_name: The OptOutListName or OptOutListArn of the OptOutList to delete. You can use DescribeOptOutLists to find the values for OptOutListName and OptOutListArn.

Main.Pinpoint_Sms_Voice_V2.delete_opted_out_number — Method

delete_opted_out_number(opt_out_list_name, opted_out_number)
delete_opted_out_number(opt_out_list_name, opted_out_number, params::Dict{String,<:Any})

Deletes an existing opted out destination phone number from the specified opt-out list. Each destination phone number can only be deleted once every 30 days. If the specified destination phone number doesn't exist or if the opt-out list doesn't exist, an Error is returned.

Arguments

opt_out_list_name: The OptOutListName or OptOutListArn to remove the phone number from.

opted_out_number: The phone number, in E.164 format, to remove from the OptOutList.

Main.Pinpoint_Sms_Voice_V2.delete_pool — Method

delete_pool(pool_id)
delete_pool(pool_id, params::Dict{String,<:Any})

Deletes an existing pool. Deleting a pool disassociates all origination identities from that pool. If the pool status isn't active or if deletion protection is enabled, an Error is returned. A pool is a collection of phone numbers and SenderIds. A pool can include one or more phone numbers and SenderIds that are associated with your Amazon Web Services account.

Arguments

pool_id: The PoolId or PoolArn of the pool to delete. You can use DescribePools to find the values for PoolId and PoolArn .

Main.Pinpoint_Sms_Voice_V2.delete_text_message_spend_limit_override — Method

delete_text_message_spend_limit_override()
delete_text_message_spend_limit_override(params::Dict{String,<:Any})

Deletes an account-level monthly spending limit override for sending text messages. Deleting a spend limit override will set the EnforcedLimit to equal the MaxLimit, which is controlled by Amazon Web Services. For more information on spend limits (quotas) see Amazon Pinpoint quotas in the Amazon Pinpoint Developer Guide.

Main.Pinpoint_Sms_Voice_V2.delete_voice_message_spend_limit_override — Method

delete_voice_message_spend_limit_override()
delete_voice_message_spend_limit_override(params::Dict{String,<:Any})

Deletes an account level monthly spend limit override for sending voice messages. Deleting a spend limit override sets the EnforcedLimit equal to the MaxLimit, which is controlled by Amazon Web Services. For more information on spending limits (quotas) see Amazon Pinpoint quotas in the Amazon Pinpoint Developer Guide.

Main.Pinpoint_Sms_Voice_V2.describe_account_attributes — Method

describe_account_attributes()
describe_account_attributes(params::Dict{String,<:Any})

Describes attributes of your Amazon Web Services account. The supported account attributes include account tier, which indicates whether your account is in the sandbox or production environment. When you're ready to move your account out of the sandbox, create an Amazon Web Services Support case for a service limit increase request. New Amazon Pinpoint accounts are placed into an SMS or voice sandbox. The sandbox protects both Amazon Web Services end recipients and SMS or voice recipients from fraud and abuse.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return per each request.

"NextToken": The token to be used for the next set of paginated results. You don't need to supply a value for this field in the initial request.

Main.Pinpoint_Sms_Voice_V2.describe_account_limits — Method

describe_account_limits()
describe_account_limits(params::Dict{String,<:Any})

Describes the current Amazon Pinpoint SMS Voice V2 resource quotas for your account. The description for a quota includes the quota name, current usage toward that quota, and the quota's maximum value. When you establish an Amazon Web Services account, the account has initial quotas on the maximum number of configuration sets, opt-out lists, phone numbers, and pools that you can create in a given Region. For more information see Amazon Pinpoint quotas in the Amazon Pinpoint Developer Guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return per each request.

"NextToken": The token to be used for the next set of paginated results. You don't need to supply a value for this field in the initial request.

Main.Pinpoint_Sms_Voice_V2.describe_configuration_sets — Method

describe_configuration_sets()
describe_configuration_sets(params::Dict{String,<:Any})

Describes the specified configuration sets or all in your account. If you specify configuration set names, the output includes information for only the specified configuration sets. If you specify filters, the output includes information for only those configuration sets that meet the filter criteria. If you don't specify configuration set names or filters, the output includes information for all configuration sets. If you specify a configuration set name that isn't valid, an error is returned.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ConfigurationSetNames": An array of strings. Each element can be either a ConfigurationSetName or ConfigurationSetArn.

"Filters": An array of filters to apply to the results that are returned.

"MaxResults": The maximum number of results to return per each request.

"NextToken": The token to be used for the next set of paginated results. You don't need to supply a value for this field in the initial request.

Main.Pinpoint_Sms_Voice_V2.describe_keywords — Method

describe_keywords(origination_identity)
describe_keywords(origination_identity, params::Dict{String,<:Any})

Describes the specified keywords or all keywords on your origination phone number or pool. A keyword is a word that you can search for on a particular phone number or pool. It is also a specific word or phrase that an end user can send to your number to elicit a response, such as an informational message or a special offer. When your number receives a message that begins with a keyword, Amazon Pinpoint responds with a customizable message. If you specify a keyword that isn't valid, an Error is returned.

Arguments

origination_identity: The origination identity to use such as a PhoneNumberId, PhoneNumberArn, SenderId or SenderIdArn. You can use DescribePhoneNumbers to find the values for PhoneNumberId and PhoneNumberArn while DescribeSenderIds can be used to get the values for SenderId and SenderIdArn.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Filters": An array of keyword filters to filter the results.

"Keywords": An array of keywords to search for.

"MaxResults": The maximum number of results to return per each request.

"NextToken": The token to be used for the next set of paginated results. You don't need to supply a value for this field in the initial request.

Main.Pinpoint_Sms_Voice_V2.describe_opt_out_lists — Method

describe_opt_out_lists()
describe_opt_out_lists(params::Dict{String,<:Any})

Describes the specified opt-out list or all opt-out lists in your account. If you specify opt-out list names, the output includes information for only the specified opt-out lists. Opt-out lists include only those that meet the filter criteria. If you don't specify opt-out list names or filters, the output includes information for all opt-out lists. If you specify an opt-out list name that isn't valid, an Error is returned.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return per each request.

"NextToken": The token to be used for the next set of paginated results. You don't need to supply a value for this field in the initial request.

"OptOutListNames": The OptOutLists to show the details of. This is an array of strings that can be either the OptOutListName or OptOutListArn.

Main.Pinpoint_Sms_Voice_V2.describe_opted_out_numbers — Method

describe_opted_out_numbers(opt_out_list_name)
describe_opted_out_numbers(opt_out_list_name, params::Dict{String,<:Any})

Describes the specified opted out destination numbers or all opted out destination numbers in an opt-out list. If you specify opted out numbers, the output includes information for only the specified opted out numbers. If you specify filters, the output includes information for only those opted out numbers that meet the filter criteria. If you don't specify opted out numbers or filters, the output includes information for all opted out destination numbers in your opt-out list. If you specify an opted out number that isn't valid, an Error is returned.

Arguments

opt_out_list_name: The OptOutListName or OptOutListArn of the OptOutList. You can use DescribeOptOutLists to find the values for OptOutListName and OptOutListArn.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Filters": An array of OptedOutFilter objects to filter the results on.

"MaxResults": The maximum number of results to return per each request.

"NextToken": The token to be used for the next set of paginated results. You don't need to supply a value for this field in the initial request.

"OptedOutNumbers": An array of phone numbers to search for in the OptOutList.

Main.Pinpoint_Sms_Voice_V2.describe_phone_numbers — Method

describe_phone_numbers()
describe_phone_numbers(params::Dict{String,<:Any})

Describes the specified origination phone number, or all the phone numbers in your account. If you specify phone number IDs, the output includes information for only the specified phone numbers. If you specify filters, the output includes information for only those phone numbers that meet the filter criteria. If you don't specify phone number IDs or filters, the output includes information for all phone numbers. If you specify a phone number ID that isn't valid, an Error is returned.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Filters": An array of PhoneNumberFilter objects to filter the results.

"MaxResults": The maximum number of results to return per each request.

"NextToken": The token to be used for the next set of paginated results. You don't need to supply a value for this field in the initial request.

"PhoneNumberIds": The unique identifier of phone numbers to find information about. This is an array of strings that can be either the PhoneNumberId or PhoneNumberArn.

Main.Pinpoint_Sms_Voice_V2.describe_pools — Method

describe_pools()
describe_pools(params::Dict{String,<:Any})

Retrieves the specified pools or all pools associated with your Amazon Web Services account. If you specify pool IDs, the output includes information for only the specified pools. If you specify filters, the output includes information for only those pools that meet the filter criteria. If you don't specify pool IDs or filters, the output includes information for all pools. If you specify a pool ID that isn't valid, an Error is returned. A pool is a collection of phone numbers and SenderIds. A pool can include one or more phone numbers and SenderIds that are associated with your Amazon Web Services account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Filters": An array of PoolFilter objects to filter the results.

"MaxResults": The maximum number of results to return per each request.

"NextToken": The token to be used for the next set of paginated results. You don't need to supply a value for this field in the initial request.

"PoolIds": The unique identifier of pools to find. This is an array of strings that can be either the PoolId or PoolArn.

Main.Pinpoint_Sms_Voice_V2.describe_sender_ids — Method

describe_sender_ids()
describe_sender_ids(params::Dict{String,<:Any})

Describes the specified SenderIds or all SenderIds associated with your Amazon Web Services account. If you specify SenderIds, the output includes information for only the specified SenderIds. If you specify filters, the output includes information for only those SenderIds that meet the filter criteria. If you don't specify SenderIds or filters, the output includes information for all SenderIds. f you specify a sender ID that isn't valid, an Error is returned.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Filters": An array of SenderIdFilter objects to filter the results.

"MaxResults": The maximum number of results to return per each request.

"NextToken": The token to be used for the next set of paginated results. You don't need to supply a value for this field in the initial request.

"SenderIds": An array of SenderIdAndCountry objects to search for.

Main.Pinpoint_Sms_Voice_V2.describe_spend_limits — Method

describe_spend_limits()
describe_spend_limits(params::Dict{String,<:Any})

Describes the current Amazon Pinpoint monthly spend limits for sending voice and text messages. When you establish an Amazon Web Services account, the account has initial monthly spend limit in a given Region. For more information on increasing your monthly spend limit, see Requesting increases to your monthly SMS spending quota for Amazon Pinpoint in the Amazon Pinpoint User Guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return per each request.

"NextToken": The token to be used for the next set of paginated results. You don't need to supply a value for this field in the initial request.

Main.Pinpoint_Sms_Voice_V2.disassociate_origination_identity — Method

disassociate_origination_identity(iso_country_code, origination_identity, pool_id)
disassociate_origination_identity(iso_country_code, origination_identity, pool_id, params::Dict{String,<:Any})

Removes the specified origination identity from an existing pool. If the origination identity isn't associated with the specified pool, an Error is returned.

Arguments

iso_country_code: The two-character code, in ISO 3166-1 alpha-2 format, for the country or region.

origination_identity: The origination identity to use such as a PhoneNumberId, PhoneNumberArn, SenderId or SenderIdArn. You can use DescribePhoneNumbers find the values for PhoneNumberId and PhoneNumberArn, or use DescribeSenderIds to get the values for SenderId and SenderIdArn.

pool_id: The unique identifier for the pool to disassociate with the origination identity. This value can be either the PoolId or PoolArn.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": Unique, case-sensitive identifier you provide to ensure the idempotency of the request. If you don't specify a client token, a randomly generated token is used for the request to ensure idempotency.

Main.Pinpoint_Sms_Voice_V2.list_pool_origination_identities — Method

list_pool_origination_identities(pool_id)
list_pool_origination_identities(pool_id, params::Dict{String,<:Any})

Lists all associated origination identities in your pool. If you specify filters, the output includes information for only those origination identities that meet the filter criteria.

Arguments

pool_id: The unique identifier for the pool. This value can be either the PoolId or PoolArn.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Filters": An array of PoolOriginationIdentitiesFilter objects to filter the results..

"MaxResults": The maximum number of results to return per each request.

"NextToken": The token to be used for the next set of paginated results. You don't need to supply a value for this field in the initial request.

Main.Pinpoint_Sms_Voice_V2.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

List all tags associated with a resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource to query for.

Main.Pinpoint_Sms_Voice_V2.put_keyword — Method

put_keyword(keyword, keyword_message, origination_identity)
put_keyword(keyword, keyword_message, origination_identity, params::Dict{String,<:Any})

Creates or updates a keyword configuration on an origination phone number or pool. A keyword is a word that you can search for on a particular phone number or pool. It is also a specific word or phrase that an end user can send to your number to elicit a response, such as an informational message or a special offer. When your number receives a message that begins with a keyword, Amazon Pinpoint responds with a customizable message. If you specify a keyword that isn't valid, an Error is returned.

Arguments

keyword: The new keyword to add.

keyword_message: The message associated with the keyword. AUTOMATICRESPONSE: A message is sent to the recipient. OPTOUT: Keeps the recipient from receiving future messages. OPT_IN: The recipient wants to receive future messages.

origination_identity: The origination identity to use such as a PhoneNumberId, PhoneNumberArn, SenderId or SenderIdArn. You can use DescribePhoneNumbers get the values for PhoneNumberId and PhoneNumberArn while DescribeSenderIds can be used to get the values for SenderId and SenderIdArn.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"KeywordAction": The action to perform for the new keyword when it is received.

Main.Pinpoint_Sms_Voice_V2.put_opted_out_number — Method

put_opted_out_number(opt_out_list_name, opted_out_number)
put_opted_out_number(opt_out_list_name, opted_out_number, params::Dict{String,<:Any})

Creates an opted out destination phone number in the opt-out list. If the destination phone number isn't valid or if the specified opt-out list doesn't exist, an Error is returned.

Arguments

opt_out_list_name: The OptOutListName or OptOutListArn to add the phone number to.

opted_out_number: The phone number to add to the OptOutList in E.164 format.

Main.Pinpoint_Sms_Voice_V2.release_phone_number — Method

release_phone_number(phone_number_id)
release_phone_number(phone_number_id, params::Dict{String,<:Any})

Releases an existing origination phone number in your account. Once released, a phone number is no longer available for sending messages. If the origination phone number has deletion protection enabled or is associated with a pool, an Error is returned.

Arguments

phone_number_id: The PhoneNumberId or PhoneNumberArn of the phone number to release. You can use DescribePhoneNumbers to get the values for PhoneNumberId and PhoneNumberArn.

Main.Pinpoint_Sms_Voice_V2.request_phone_number — Method

request_phone_number(iso_country_code, message_type, number_capabilities, number_type)
request_phone_number(iso_country_code, message_type, number_capabilities, number_type, params::Dict{String,<:Any})

Request an origination phone number for use in your account. For more information on phone number request see Requesting a number in the Amazon Pinpoint User Guide.

Arguments

iso_country_code: The two-character code, in ISO 3166-1 alpha-2 format, for the country or region.

message_type: The type of message. Valid values are TRANSACTIONAL for messages that are critical or time-sensitive and PROMOTIONAL for messages that aren't critical or time-sensitive.

number_capabilities: Indicates if the phone number will be used for text messages, voice messages, or both.

number_type: The type of phone number to request.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If you don't specify a client token, a randomly generated token is used for the request to ensure idempotency.

"DeletionProtectionEnabled": By default this is set to false. When set to true the phone number can't be deleted.

"OptOutListName": The name of the OptOutList to associate with the phone number. You can use the OutOutListName or OptPutListArn.

"PoolId": The pool to associated with the phone number. You can use the PoolId or PoolArn.

"RegistrationId": Use this field to attach your phone number for an external registration process.

"Tags": An array of tags (key and value pairs) associate with the requested phone number.

Main.Pinpoint_Sms_Voice_V2.send_text_message — Method

send_text_message(destination_phone_number)
send_text_message(destination_phone_number, params::Dict{String,<:Any})

Creates a new text message and sends it to a recipient's phone number. SMS throughput limits are measured in Message Parts per Second (MPS). Your MPS limit depends on the destination country of your messages, as well as the type of phone number (origination number) that you use to send the message. For more information, see Message Parts per Second (MPS) limits in the Amazon Pinpoint User Guide.

Arguments

destination_phone_number: The destination phone number in E.164 format.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ConfigurationSetName": The name of the configuration set to use. This can be either the ConfigurationSetName or ConfigurationSetArn.

"Context": You can specify custom data in this field. If you do, that data is logged to the event destination.

"DestinationCountryParameters": This field is used for any country-specific registration requirements. Currently, this setting is only used when you send messages to recipients in India using a sender ID. For more information see Special requirements for sending SMS messages to recipients in India.

"DryRun": When set to true, the message is checked and validated, but isn't sent to the end recipient.

"Keyword": When you register a short code in the US, you must specify a program name. If you don’t have a US short code, omit this attribute.

"MaxPrice": The maximum amount that you want to spend, in US dollars, per each text message part. A text message can contain multiple parts.

"MessageBody": The body of the text message.

"MessageType": The type of message. Valid values are TRANSACTIONAL for messages that are critical or time-sensitive and PROMOTIONAL for messages that aren't critical or time-sensitive.

"OriginationIdentity": The origination identity of the message. This can be either the PhoneNumber, PhoneNumberId, PhoneNumberArn, SenderId, SenderIdArn, PoolId, or PoolArn.

"TimeToLive": How long the text message is valid for. By default this is 72 hours.

Main.Pinpoint_Sms_Voice_V2.send_voice_message — Method

send_voice_message(destination_phone_number, origination_identity)
send_voice_message(destination_phone_number, origination_identity, params::Dict{String,<:Any})

Allows you to send a request that sends a text message through Amazon Pinpoint. This operation uses Amazon Polly to convert a text script into a voice message.

Arguments

destination_phone_number: The destination phone number in E.164 format.

origination_identity: The origination identity to use for the voice call. This can be the PhoneNumber, PhoneNumberId, PhoneNumberArn, PoolId, or PoolArn.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ConfigurationSetName": The name of the configuration set to use. This can be either the ConfigurationSetName or ConfigurationSetArn.

"Context": You can specify custom data in this field. If you do, that data is logged to the event destination.

"DryRun": When set to true, the message is checked and validated, but isn't sent to the end recipient.

"MaxPricePerMinute": The maximum amount to spend per voice message, in US dollars.

"MessageBody": The text to convert to a voice message.

"MessageBodyTextType": Specifies if the MessageBody field contains text or speech synthesis markup language (SSML). TEXT: This is the default value. When used the maximum character limit is 3000. SSML: When used the maximum character limit is 6000 including SSML tagging.

"TimeToLive": How long the voice message is valid for. By default this is 72 hours.

"VoiceId": The voice for the Amazon Polly service to use. By default this is set to "MATTHEW".

Main.Pinpoint_Sms_Voice_V2.set_default_message_type — Method

set_default_message_type(configuration_set_name, message_type)
set_default_message_type(configuration_set_name, message_type, params::Dict{String,<:Any})

Sets the default message type on a configuration set. Choose the category of SMS messages that you plan to send from this account. If you send account-related messages or time-sensitive messages such as one-time passcodes, choose Transactional. If you plan to send messages that contain marketing material or other promotional content, choose Promotional. This setting applies to your entire Amazon Web Services account.

Arguments

configuration_set_name: The configuration set to update with a new default message type. This field can be the ConsigurationSetName or ConfigurationSetArn.

message_type: The type of message. Valid values are TRANSACTIONAL for messages that are critical or time-sensitive and PROMOTIONAL for messages that aren't critical or time-sensitive.

Main.Pinpoint_Sms_Voice_V2.set_default_sender_id — Method

set_default_sender_id(configuration_set_name, sender_id)
set_default_sender_id(configuration_set_name, sender_id, params::Dict{String,<:Any})

Sets default sender ID on a configuration set. When sending a text message to a destination country that supports sender IDs, the default sender ID on the configuration set specified will be used if no dedicated origination phone numbers or registered sender IDs are available in your account.

Arguments

configuration_set_name: The configuration set to updated with a new default SenderId. This field can be the ConsigurationSetName or ConfigurationSetArn.

sender_id: The current sender ID for the configuration set. When sending a text message to a destination country which supports SenderIds, the default sender ID on the configuration set specified on SendTextMessage will be used if no dedicated origination phone numbers or registered SenderIds are available in your account, instead of a generic sender ID, such as 'NOTICE'.

Main.Pinpoint_Sms_Voice_V2.set_text_message_spend_limit_override — Method

set_text_message_spend_limit_override(monthly_limit)
set_text_message_spend_limit_override(monthly_limit, params::Dict{String,<:Any})

Sets an account level monthly spend limit override for sending text messages. The requested spend limit must be less than or equal to the MaxLimit, which is set by Amazon Web Services.

Arguments

monthly_limit: The new monthly limit to enforce on text messages.

Main.Pinpoint_Sms_Voice_V2.set_voice_message_spend_limit_override — Method

set_voice_message_spend_limit_override(monthly_limit)
set_voice_message_spend_limit_override(monthly_limit, params::Dict{String,<:Any})

Sets an account level monthly spend limit override for sending voice messages. The requested spend limit must be less than or equal to the MaxLimit, which is set by Amazon Web Services.

Arguments

monthly_limit: The new monthly limit to enforce on voice messages.

Main.Pinpoint_Sms_Voice_V2.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Adds or overwrites only the specified tags for the specified Amazon Pinpoint SMS Voice, version 2 resource. When you specify an existing tag key, the value is overwritten with the new value. Each resource can have a maximum of 50 tags. Each tag consists of a key and an optional value. Tag keys must be unique per resource. For more information about tags, see Tagging Amazon Pinpoint resources in the Amazon Pinpoint Developer Guide.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

tags: An array of key and value pair tags that are associated with the resource.

Main.Pinpoint_Sms_Voice_V2.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes the association of the specified tags from an Amazon Pinpoint SMS Voice V2 resource. For more information on tags see Tagging Amazon Pinpoint resources in the Amazon Pinpoint Developer Guide.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

tag_keys: An array of tag key values to unassociate with the resource.

Main.Pinpoint_Sms_Voice_V2.update_event_destination — Method

update_event_destination(configuration_set_name, event_destination_name)
update_event_destination(configuration_set_name, event_destination_name, params::Dict{String,<:Any})

Updates an existing event destination in a configuration set. You can update the IAM role ARN for CloudWatch Logs and Kinesis Data Firehose. You can also enable or disable the event destination. You may want to update an event destination to change its matching event types or updating the destination resource ARN. You can't change an event destination's type between CloudWatch Logs, Kinesis Data Firehose, and Amazon SNS.

Arguments

configuration_set_name: The configuration set to update with the new event destination. Valid values for this can be the ConfigurationSetName or ConfigurationSetArn.

event_destination_name: The name to use for the event destination.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"CloudWatchLogsDestination": An object that contains information about an event destination that sends data to CloudWatch Logs.

"Enabled": When set to true logging is enabled.

"KinesisFirehoseDestination": An object that contains information about an event destination for logging to Kinesis Data Firehose.

"MatchingEventTypes": An array of event types that determine which events to log.

"SnsDestination": An object that contains information about an event destination that sends data to Amazon SNS.

Main.Pinpoint_Sms_Voice_V2.update_phone_number — Method

update_phone_number(phone_number_id)
update_phone_number(phone_number_id, params::Dict{String,<:Any})

Updates the configuration of an existing origination phone number. You can update the opt-out list, enable or disable two-way messaging, change the TwoWayChannelArn, enable or disable self-managed opt-outs, and enable or disable deletion protection. If the origination phone number is associated with a pool, an Error is returned.

Arguments

phone_number_id: The unique identifier of the phone number. Valid values for this field can be either the PhoneNumberId or PhoneNumberArn.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DeletionProtectionEnabled": By default this is set to false. When set to true the phone number can't be deleted.

"OptOutListName": The OptOutList to add the phone number to. Valid values for this field can be either the OutOutListName or OutOutListArn.

"SelfManagedOptOutsEnabled": By default this is set to false. When an end recipient sends a message that begins with HELP or STOP to one of your dedicated numbers, Amazon Pinpoint automatically replies with a customizable message and adds the end recipient to the OptOutList. When set to true you're responsible for responding to HELP and STOP requests. You're also responsible for tracking and honoring opt-out requests.

"TwoWayChannelArn": The Amazon Resource Name (ARN) of the two way channel.

"TwoWayEnabled": By default this is set to false. When set to true you can receive incoming text messages from your end recipients.

Main.Pinpoint_Sms_Voice_V2.update_pool — Method

update_pool(pool_id)
update_pool(pool_id, params::Dict{String,<:Any})

Updates the configuration of an existing pool. You can update the opt-out list, enable or disable two-way messaging, change the TwoWayChannelArn, enable or disable self-managed opt-outs, enable or disable deletion protection, and enable or disable shared routes.

Arguments

pool_id: The unique identifier of the pool to update. Valid values are either the PoolId or PoolArn.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DeletionProtectionEnabled": When set to true the pool can't be deleted.

"OptOutListName": The OptOutList to associate with the pool. Valid values are either OptOutListName or OptOutListArn.

"SelfManagedOptOutsEnabled": By default this is set to false. When an end recipient sends a message that begins with HELP or STOP to one of your dedicated numbers, Amazon Pinpoint automatically replies with a customizable message and adds the end recipient to the OptOutList. When set to true you're responsible for responding to HELP and STOP requests. You're also responsible for tracking and honoring opt-out requests.

"SharedRoutesEnabled": Indicates whether shared routes are enabled for the pool.

"TwoWayChannelArn": The Amazon Resource Name (ARN) of the two way channel.

"TwoWayEnabled": By default this is set to false. When set to true you can receive incoming text messages from your end recipients.