Privatenetworks

This page documents function available when using the Privatenetworks module, created with @service Privatenetworks.

Index

Main.Privatenetworks.acknowledge_order_receipt

Main.Privatenetworks.activate_device_identifier

Main.Privatenetworks.activate_network_site

Main.Privatenetworks.configure_access_point

Main.Privatenetworks.create_network

Main.Privatenetworks.create_network_site

Main.Privatenetworks.deactivate_device_identifier

Main.Privatenetworks.delete_network

Main.Privatenetworks.delete_network_site

Main.Privatenetworks.get_device_identifier

Main.Privatenetworks.get_network

Main.Privatenetworks.get_network_resource

Main.Privatenetworks.get_network_site

Main.Privatenetworks.get_order

Main.Privatenetworks.list_device_identifiers

Main.Privatenetworks.list_network_resources

Main.Privatenetworks.list_network_sites

Main.Privatenetworks.list_networks

Main.Privatenetworks.list_orders

Main.Privatenetworks.list_tags_for_resource

Main.Privatenetworks.ping

Main.Privatenetworks.start_network_resource_update

Main.Privatenetworks.tag_resource

Main.Privatenetworks.untag_resource

Main.Privatenetworks.update_network_site

Main.Privatenetworks.update_network_site_plan

Documentation

Main.Privatenetworks.acknowledge_order_receipt — Method

acknowledge_order_receipt(order_arn)
acknowledge_order_receipt(order_arn, params::Dict{String,<:Any})

Acknowledges that the specified network order was received.

Arguments

order_arn: The Amazon Resource Name (ARN) of the order.

Main.Privatenetworks.activate_device_identifier — Method

activate_device_identifier(device_identifier_arn)
activate_device_identifier(device_identifier_arn, params::Dict{String,<:Any})

Activates the specified device identifier.

Arguments

device_identifier_arn: The Amazon Resource Name (ARN) of the device identifier.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

Main.Privatenetworks.activate_network_site — Method

activate_network_site(network_site_arn, shipping_address)
activate_network_site(network_site_arn, shipping_address, params::Dict{String,<:Any})

Activates the specified network site.

Arguments

network_site_arn: The Amazon Resource Name (ARN) of the network site.

shipping_address: The shipping address of the network site.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

Main.Privatenetworks.configure_access_point — Method

configure_access_point(access_point_arn)
configure_access_point(access_point_arn, params::Dict{String,<:Any})

Configures the specified network resource. Use this action to specify the geographic position of the hardware. You must provide Certified Professional Installer (CPI) credentials in the request so that we can obtain spectrum grants. For more information, see Radio units in the Amazon Web Services Private 5G User Guide.

Arguments

access_point_arn: The Amazon Resource Name (ARN) of the network resource.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"cpiSecretKey": A Base64 encoded string of the CPI certificate associated with the CPI user who is certifying the coordinates of the network resource.

"cpiUserId": The CPI user ID of the CPI user who is certifying the coordinates of the network resource.

"cpiUserPassword": The CPI password associated with the CPI certificate in cpiSecretKey.

"cpiUsername": The CPI user name of the CPI user who is certifying the coordinates of the radio unit.

"position": The position of the network resource.

Main.Privatenetworks.create_network — Method

create_network(network_name)
create_network(network_name, params::Dict{String,<:Any})

Creates a network.

Arguments

network_name: The name of the network. You can't change the name after you create the network.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

"description": The description of the network.

"tags": The tags to apply to the network.

Main.Privatenetworks.create_network_site — Method

create_network_site(network_arn, network_site_name)
create_network_site(network_arn, network_site_name, params::Dict{String,<:Any})

Creates a network site.

Arguments

network_arn: The Amazon Resource Name (ARN) of the network.

network_site_name: The name of the site. You can't change the name after you create the site.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"availabilityZone": The Availability Zone that is the parent of this site. You can't change the Availability Zone after you create the site.

"availabilityZoneId": The ID of the Availability Zone that is the parent of this site. You can't change the Availability Zone after you create the site.

"clientToken": Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

"description": The description of the site.

"pendingPlan": Information about the pending plan for this site.

"tags": The tags to apply to the network site.

Main.Privatenetworks.deactivate_device_identifier — Method

deactivate_device_identifier(device_identifier_arn)
deactivate_device_identifier(device_identifier_arn, params::Dict{String,<:Any})

Deactivates the specified device identifier.

Arguments

device_identifier_arn: The Amazon Resource Name (ARN) of the device identifier.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

Main.Privatenetworks.delete_network — Method

delete_network(network_arn)
delete_network(network_arn, params::Dict{String,<:Any})

Deletes the specified network. You must delete network sites before you delete the network. For more information, see DeleteNetworkSite in the API Reference for Amazon Web Services Private 5G.

Arguments

network_arn: The Amazon Resource Name (ARN) of the network.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

Main.Privatenetworks.delete_network_site — Method

delete_network_site(network_site_arn)
delete_network_site(network_site_arn, params::Dict{String,<:Any})

Deletes the specified network site. Return the hardware after you delete the network site. You are responsible for minimum charges. For more information, see Hardware returns in the Amazon Web Services Private 5G User Guide.

Arguments

network_site_arn: The Amazon Resource Name (ARN) of the network site.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

Main.Privatenetworks.get_device_identifier — Method

get_device_identifier(device_identifier_arn)
get_device_identifier(device_identifier_arn, params::Dict{String,<:Any})

Gets the specified device identifier.

Arguments

device_identifier_arn: The Amazon Resource Name (ARN) of the device identifier.

Main.Privatenetworks.get_network — Method

get_network(network_arn)
get_network(network_arn, params::Dict{String,<:Any})

Gets the specified network.

Arguments

network_arn: The Amazon Resource Name (ARN) of the network.

Main.Privatenetworks.get_network_resource — Method

get_network_resource(network_resource_arn)
get_network_resource(network_resource_arn, params::Dict{String,<:Any})

Gets the specified network resource.

Arguments

network_resource_arn: The Amazon Resource Name (ARN) of the network resource.

Main.Privatenetworks.get_network_site — Method

get_network_site(network_site_arn)
get_network_site(network_site_arn, params::Dict{String,<:Any})

Gets the specified network site.

Arguments

network_site_arn: The Amazon Resource Name (ARN) of the network site.

Main.Privatenetworks.get_order — Method

get_order(order_arn)
get_order(order_arn, params::Dict{String,<:Any})

Gets the specified order.

Arguments

order_arn: The Amazon Resource Name (ARN) of the order.

Main.Privatenetworks.list_device_identifiers — Method

list_device_identifiers(network_arn)
list_device_identifiers(network_arn, params::Dict{String,<:Any})

Lists device identifiers. Add filters to your request to return a more specific list of results. Use filters to match the Amazon Resource Name (ARN) of an order, the status of device identifiers, or the ARN of the traffic group. If you specify multiple filters, filters are joined with an OR, and the request returns results that match all of the specified filters.

Arguments

network_arn: The Amazon Resource Name (ARN) of the network.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filters": The filters. ORDER - The Amazon Resource Name (ARN) of the order. STATUS - The status (ACTIVE | INACTIVE). TRAFFIC_GROUP - The Amazon Resource Name (ARN) of the traffic group. Filter values are case sensitive. If you specify multiple values for a filter, the values are joined with an OR, and the request returns all results that match any of the specified values.

"maxResults": The maximum number of results to return.

"startToken": The token for the next page of results.

Main.Privatenetworks.list_network_resources — Method

list_network_resources(network_arn)
list_network_resources(network_arn, params::Dict{String,<:Any})

Lists network resources. Add filters to your request to return a more specific list of results. Use filters to match the Amazon Resource Name (ARN) of an order or the status of network resources. If you specify multiple filters, filters are joined with an OR, and the request returns results that match all of the specified filters.

Arguments

network_arn: The Amazon Resource Name (ARN) of the network.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filters": The filters. ORDER - The Amazon Resource Name (ARN) of the order. STATUS - The status (AVAILABLE | DELETED | DELETING | PENDING | PENDING_RETURN | PROVISIONING | SHIPPED). Filter values are case sensitive. If you specify multiple values for a filter, the values are joined with an OR, and the request returns all results that match any of the specified values.

"maxResults": The maximum number of results to return.

"startToken": The token for the next page of results.

Main.Privatenetworks.list_network_sites — Method

list_network_sites(network_arn)
list_network_sites(network_arn, params::Dict{String,<:Any})

Lists network sites. Add filters to your request to return a more specific list of results. Use filters to match the status of the network site.

Arguments

network_arn: The Amazon Resource Name (ARN) of the network.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filters": The filters. Add filters to your request to return a more specific list of results. Use filters to match the status of the network sites. STATUS - The status (AVAILABLE | CREATED | DELETED | DEPROVISIONING | PROVISIONING). Filter values are case sensitive. If you specify multiple values for a filter, the values are joined with an OR, and the request returns all results that match any of the specified values.

"maxResults": The maximum number of results to return.

"startToken": The token for the next page of results.

Main.Privatenetworks.list_networks — Method

list_networks()
list_networks(params::Dict{String,<:Any})

Lists networks. Add filters to your request to return a more specific list of results. Use filters to match the status of the network.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filters": The filters. STATUS - The status (AVAILABLE | CREATED | DELETED | DEPROVISIONING | PROVISIONING). Filter values are case sensitive. If you specify multiple values for a filter, the values are joined with an OR, and the request returns all results that match any of the specified values.

"maxResults": The maximum number of results to return.

"startToken": The token for the next page of results.

Main.Privatenetworks.list_orders — Method

list_orders(network_arn)
list_orders(network_arn, params::Dict{String,<:Any})

Lists orders. Add filters to your request to return a more specific list of results. Use filters to match the Amazon Resource Name (ARN) of the network site or the status of the order. If you specify multiple filters, filters are joined with an OR, and the request returns results that match all of the specified filters.

Arguments

network_arn: The Amazon Resource Name (ARN) of the network.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filters": The filters. NETWORK_SITE - The Amazon Resource Name (ARN) of the network site. STATUS - The status (ACKNOWLEDGING | ACKNOWLEDGED | UNACKNOWLEDGED). Filter values are case sensitive. If you specify multiple values for a filter, the values are joined with an OR, and the request returns all results that match any of the specified values.

"maxResults": The maximum number of results to return.

"startToken": The token for the next page of results.

Main.Privatenetworks.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists the tags for the specified resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

Main.Privatenetworks.ping — Method

ping()
ping(params::Dict{String,<:Any})

Checks the health of the service.

Main.Privatenetworks.start_network_resource_update — Method

start_network_resource_update(network_resource_arn, update_type)
start_network_resource_update(network_resource_arn, update_type, params::Dict{String,<:Any})

Starts an update of the specified network resource. After you submit a request to replace or return a network resource, the status of the network resource is CREATINGSHIPPINGLABEL. The shipping label is available when the status of the network resource is PENDING_RETURN. After the network resource is successfully returned, its status is DELETED. For more information, see Return a radio unit.

Arguments

network_resource_arn: The Amazon Resource Name (ARN) of the network resource.

update_type: The update type. REPLACE - Submits a request to replace a defective radio unit. We provide a shipping label that you can use for the return process and we ship a replacement radio unit to you. RETURN - Submits a request to replace a radio unit that you no longer need. We provide a shipping label that you can use for the return process.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"returnReason": The reason for the return. Providing a reason for a return is optional.

"shippingAddress": The shipping address. If you don't provide a shipping address when replacing or returning a network resource, we use the address from the original order for the network resource.

Main.Privatenetworks.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Adds tags to the specified resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

tags: The tags to add to the resource.

Main.Privatenetworks.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes tags from the specified resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

tag_keys: The tag keys.

Main.Privatenetworks.update_network_site — Method

update_network_site(network_site_arn)
update_network_site(network_site_arn, params::Dict{String,<:Any})

Updates the specified network site.

Arguments

network_site_arn: The Amazon Resource Name (ARN) of the network site.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.

"description": The description.

Main.Privatenetworks.update_network_site_plan — Method

update_network_site_plan(network_site_arn, pending_plan)
update_network_site_plan(network_site_arn, pending_plan, params::Dict{String,<:Any})

Updates the specified network site plan.

Arguments

network_site_arn: The Amazon Resource Name (ARN) of the network site.

pending_plan: The pending plan.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Unique, case-sensitive identifier that you provide to ensure the idempotency of the request. For more information, see How to ensure idempotency.