Proton

This page documents function available when using the Proton module, created with @service Proton.

Index

Main.Proton.accept_environment_account_connection

Main.Proton.cancel_environment_deployment

Main.Proton.cancel_service_instance_deployment

Main.Proton.cancel_service_pipeline_deployment

Main.Proton.create_environment

Main.Proton.create_environment_account_connection

Main.Proton.create_environment_template

Main.Proton.create_environment_template_version

Main.Proton.create_service

Main.Proton.create_service_template

Main.Proton.create_service_template_version

Main.Proton.delete_environment

Main.Proton.delete_environment_account_connection

Main.Proton.delete_environment_template

Main.Proton.delete_environment_template_version

Main.Proton.delete_service

Main.Proton.delete_service_template

Main.Proton.delete_service_template_version

Main.Proton.get_account_settings

Main.Proton.get_environment

Main.Proton.get_environment_account_connection

Main.Proton.get_environment_template

Main.Proton.get_environment_template_version

Main.Proton.get_service

Main.Proton.get_service_instance

Main.Proton.get_service_template

Main.Proton.get_service_template_version

Main.Proton.list_environment_account_connections

Main.Proton.list_environment_template_versions

Main.Proton.list_environment_templates

Main.Proton.list_environments

Main.Proton.list_service_instances

Main.Proton.list_service_template_versions

Main.Proton.list_service_templates

Main.Proton.list_services

Main.Proton.list_tags_for_resource

Main.Proton.reject_environment_account_connection

Main.Proton.tag_resource

Main.Proton.untag_resource

Main.Proton.update_account_settings

Main.Proton.update_environment

Main.Proton.update_environment_account_connection

Main.Proton.update_environment_template

Main.Proton.update_environment_template_version

Main.Proton.update_service

Main.Proton.update_service_instance

Main.Proton.update_service_pipeline

Main.Proton.update_service_template

Main.Proton.update_service_template_version

Documentation

Main.Proton.accept_environment_account_connection — Method

accept_environment_account_connection(id)
accept_environment_account_connection(id, params::Dict{String,<:Any})

In a management account, an environment account connection request is accepted. When the environment account connection request is accepted, AWS Proton can use the associated IAM role to provision environment infrastructure resources in the associated environment account. For more information, see Environment account connections in the AWS Proton Administration guide.

Arguments

id: The ID of the environment account connection.

Main.Proton.cancel_environment_deployment — Method

cancel_environment_deployment(environment_name)
cancel_environment_deployment(environment_name, params::Dict{String,<:Any})

Attempts to cancel an environment deployment on an UpdateEnvironment action, if the deployment is IN_PROGRESS. For more information, see Update an environment in the AWS Proton Administration guide. The following list includes potential cancellation scenarios. If the cancellation attempt succeeds, the resulting deployment state is CANCELLED. If the cancellation attempt fails, the resulting deployment state is FAILED. If the current UpdateEnvironment action succeeds before the cancellation attempt starts, the resulting deployment state is SUCCEEDED and the cancellation attempt has no effect.

Arguments

environment_name: The name of the environment with the deployment to cancel.

Main.Proton.cancel_service_instance_deployment — Method

cancel_service_instance_deployment(service_instance_name, service_name)
cancel_service_instance_deployment(service_instance_name, service_name, params::Dict{String,<:Any})

Attempts to cancel a service instance deployment on an UpdateServiceInstance action, if the deployment is IN_PROGRESS. For more information, see Update a service instance in the AWS Proton Administration guide or the AWS Proton User guide. The following list includes potential cancellation scenarios. If the cancellation attempt succeeds, the resulting deployment state is CANCELLED. If the cancellation attempt fails, the resulting deployment state is FAILED. If the current UpdateServiceInstance action succeeds before the cancellation attempt starts, the resulting deployment state is SUCCEEDED and the cancellation attempt has no effect.

Arguments

service_instance_name: The name of the service instance with the deployment to cancel.

service_name: The name of the service with the service instance deployment to cancel.

Main.Proton.cancel_service_pipeline_deployment — Method

cancel_service_pipeline_deployment(service_name)
cancel_service_pipeline_deployment(service_name, params::Dict{String,<:Any})

Attempts to cancel a service pipeline deployment on an UpdateServicePipeline action, if the deployment is IN_PROGRESS. For more information, see Update a service pipeline in the AWS Proton Administration guide or the AWS Proton User guide. The following list includes potential cancellation scenarios. If the cancellation attempt succeeds, the resulting deployment state is CANCELLED. If the cancellation attempt fails, the resulting deployment state is FAILED. If the current UpdateServicePipeline action succeeds before the cancellation attempt starts, the resulting deployment state is SUCCEEDED and the cancellation attempt has no effect.

Arguments

service_name: The name of the service with the service pipeline deployment to cancel.

Main.Proton.create_environment — Method

create_environment(name, spec, template_major_version, template_name)
create_environment(name, spec, template_major_version, template_name, params::Dict{String,<:Any})

Deploy a new environment. An AWS Proton environment is created from an environment template that defines infrastructure and resources that can be shared across services. For more information, see the Environments in the AWS Proton Administration Guide.

Arguments

name: The name of the environment.

spec: A link to a YAML formatted spec file that provides inputs as defined in the environment template bundle schema file. For more information, see Environments in the AWS Proton Administration Guide.

template_major_version: The ID of the major version of the environment template.

template_name: The name of the environment template. For more information, see Environment Templates in the AWS Proton Administration Guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A description of the environment that's being created and deployed.

"environmentAccountConnectionId": The ID of the environment account connection that you provide if you're provisioning your environment infrastructure resources to an environment account. You must include either the environmentAccountConnectionId or protonServiceRoleArn parameter and value. For more information, see Environment account connections in the AWS Proton Administration guide.

"protonServiceRoleArn": The Amazon Resource Name (ARN) of the AWS Proton service role that allows AWS Proton to make calls to other services on your behalf. You must include either the environmentAccountConnectionId or protonServiceRoleArn parameter and value.

"tags": Create tags for your environment. For more information, see AWS Proton resources and tagging in the AWS Proton Administration Guide or AWS Proton User Guide.

"templateMinorVersion": The ID of the minor version of the environment template.

Main.Proton.create_environment_account_connection — Method

create_environment_account_connection(environment_name, management_account_id, role_arn)
create_environment_account_connection(environment_name, management_account_id, role_arn, params::Dict{String,<:Any})

Create an environment account connection in an environment account so that environment infrastructure resources can be provisioned in the environment account from the management account. An environment account connection is a secure bi-directional connection between a management account and an environment account that maintains authorization and permissions. For more information, see Environment account connections in the AWS Proton Administration guide.

Arguments

environment_name: The name of the AWS Proton environment that's created in the associated management account.

management_account_id: The ID of the management account that accepts or rejects the environment account connection. You create an manage the AWS Proton environment in this account. If the management account accepts the environment account connection, AWS Proton can use the associated IAM role to provision environment infrastructure resources in the associated environment account.

role_arn: The Amazon Resource Name (ARN) of the IAM service role that's created in the environment account. AWS Proton uses this role to provision infrastructure resources in the associated environment account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": When included, if two identicial requests are made with the same client token, AWS Proton returns the environment account connection that the first request created.

Main.Proton.create_environment_template — Method

create_environment_template(name)
create_environment_template(name, params::Dict{String,<:Any})

Create an environment template for AWS Proton. For more information, see Environment Templates in the AWS Proton Administration Guide. You can create an environment template in one of the two following ways: Register and publish a standard environment template that instructs AWS Proton to deploy and manage environment infrastructure. Register and publish a customer managed environment template that connects AWS Proton to your existing provisioned infrastructure that you manage. AWS Proton doesn't manage your existing provisioned infrastructure. To create an environment template for customer provisioned and managed infrastructure, include the provisioning parameter and set the value to CUSTOMER_MANAGED. For more information, see Register and publish an environment template in the AWS Proton Administration Guide.

Arguments

name: The name of the environment template.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A description of the environment template.

"displayName": The environment template name as displayed in the developer interface.

"encryptionKey": A customer provided encryption key that AWS Proton uses to encrypt data.

"provisioning": When included, indicates that the environment template is for customer provisioned and managed infrastructure.

"tags": Create tags for your environment template. For more information, see AWS Proton resources and tagging in the AWS Proton Administration Guide or AWS Proton User Guide.

Main.Proton.create_environment_template_version — Method

create_environment_template_version(source, template_name)
create_environment_template_version(source, template_name, params::Dict{String,<:Any})

Create a new major or minor version of an environment template. A major version of an environment template is a version that isn't backwards compatible. A minor version of an environment template is a version that's backwards compatible within its major version.

Arguments

source: An object that includes the template bundle S3 bucket path and name for the new version of an template.

template_name: The name of the environment template.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": When included, if two identicial requests are made with the same client token, AWS Proton returns the environment template version that the first request created.

"description": A description of the new version of an environment template.

"majorVersion": To create a new minor version of the environment template, include a majorVersion. To create a new major and minor version of the environment template, exclude majorVersion.

"tags": Create tags for a new version of an environment template.

Main.Proton.create_service — Method

create_service(name, spec, template_major_version, template_name)
create_service(name, spec, template_major_version, template_name, params::Dict{String,<:Any})

Create an AWS Proton service. An AWS Proton service is an instantiation of a service template and often includes several service instances and pipeline. For more information, see Services in the AWS Proton Administration Guide and Services in the AWS Proton User Guide.

Arguments

name: The service name.

spec: A link to a spec file that provides inputs as defined in the service template bundle schema file. The spec file is in YAML format. Don’t include pipeline inputs in the spec if your service template doesn’t include a service pipeline. For more information, see Create a service in the AWS Proton Administration Guide and Create a service in the AWS Proton User Guide.

template_major_version: The ID of the major version of the service template that was used to create the service.

template_name: The name of the service template that's used to create the service.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"branchName": The name of the code repository branch that holds the code that's deployed in AWS Proton. Don't include this parameter if your service template doesn't include a service pipeline.

"description": A description of the AWS Proton service.

"repositoryConnectionArn": The ARN of the repository connection. For more information, see Set up repository connection in the AWS Proton Administration Guide and Getting started in the AWS Proton User Guide. Don't include this parameter if your service template doesn't include a service pipeline.

"repositoryId": The ID of the code repository. Don't include this parameter if your service template doesn't include a service pipeline.

"tags": Create tags for your service. For more information, see AWS Proton resources and tagging in the AWS Proton Administration Guide or AWS Proton User Guide.

"templateMinorVersion": The ID of the minor version of the service template that was used to create the service.

Main.Proton.create_service_template — Method

create_service_template(name)
create_service_template(name, params::Dict{String,<:Any})

Create a service template. The administrator creates a service template to define standardized infrastructure and an optional CICD service pipeline. Developers, in turn, select the service template from AWS Proton. If the selected service template includes a service pipeline definition, they provide a link to their source code repository. AWS Proton then deploys and manages the infrastructure defined by the selected service template. For more information, see Service Templates in the AWS Proton Administration Guide.

Arguments

name: The name of the service template.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A description of the service template.

"displayName": The name of the service template as displayed in the developer interface.

"encryptionKey": A customer provided encryption key that's used to encrypt data.

"pipelineProvisioning": AWS Proton includes a service pipeline for your service by default. When included, this parameter indicates that an AWS Proton service pipeline won't be included for your service. Once specified, this parameter can't be changed. For more information, see Service template bundles in the AWS Proton Administration Guide.

"tags": Create tags for your service template. For more information, see AWS Proton resources and tagging in the AWS Proton Administration Guide or AWS Proton User Guide.

Main.Proton.create_service_template_version — Method

create_service_template_version(compatible_environment_templates, source, template_name)
create_service_template_version(compatible_environment_templates, source, template_name, params::Dict{String,<:Any})

Create a new major or minor version of a service template. A major version of a service template is a version that isn't backwards compatible. A minor version of a service template is a version that's backwards compatible within its major version.

Arguments

compatible_environment_templates: An array of compatible environment template objects for the new version of a service template.

source: An object that includes the template bundle S3 bucket path and name for the new version of a service template.

template_name: The name of the service template.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": When included, if two identicial requests are made with the same client token, AWS Proton returns the service template version that the first request created.

"description": A description of the new version of a service template.

"majorVersion": To create a new minor version of the service template, include a majorVersion. To create a new major and minor version of the service template, exclude majorVersion.

"tags": Create tags for a new version of a service template.

Main.Proton.delete_environment — Method

delete_environment(name)
delete_environment(name, params::Dict{String,<:Any})

Delete an environment.

Arguments

name: The name of the environment to delete.

Main.Proton.delete_environment_account_connection — Method

delete_environment_account_connection(id)
delete_environment_account_connection(id, params::Dict{String,<:Any})

In an environment account, delete an environment account connection. After you delete an environment account connection that’s in use by an AWS Proton environment, AWS Proton can’t manage the environment infrastructure resources until a new environment account connection is accepted for the environment account and associated environment. You're responsible for cleaning up provisioned resources that remain without an environment connection. For more information, see Environment account connections in the AWS Proton Administration guide.

Arguments

id: The ID of the environment account connection to delete.

Main.Proton.delete_environment_template — Method

delete_environment_template(name)
delete_environment_template(name, params::Dict{String,<:Any})

If no other major or minor versions of an environment template exist, delete the environment template.

Arguments

name: The name of the environment template to delete.

Main.Proton.delete_environment_template_version — Method

delete_environment_template_version(major_version, minor_version, template_name)
delete_environment_template_version(major_version, minor_version, template_name, params::Dict{String,<:Any})

If no other minor versions of an environment template exist, delete a major version of the environment template if it's not the Recommended version. Delete the Recommended version of the environment template if no other major versions or minor versions of the environment template exist. A major version of an environment template is a version that's not backwards compatible. Delete a minor version of an environment template if it isn't the Recommended version. Delete a Recommended minor version of the environment template if no other minor versions of the environment template exist. A minor version of an environment template is a version that's backwards compatible.

Arguments

major_version: The environment template major version to delete.

minor_version: The environment template minor version to delete.

template_name: The name of the environment template.

Main.Proton.delete_service — Method

delete_service(name)
delete_service(name, params::Dict{String,<:Any})

Delete a service.

Arguments

name: The name of the service to delete.

Main.Proton.delete_service_template — Method

delete_service_template(name)
delete_service_template(name, params::Dict{String,<:Any})

If no other major or minor versions of the service template exist, delete the service template.

Arguments

name: The name of the service template to delete.

Main.Proton.delete_service_template_version — Method

delete_service_template_version(major_version, minor_version, template_name)
delete_service_template_version(major_version, minor_version, template_name, params::Dict{String,<:Any})

If no other minor versions of a service template exist, delete a major version of the service template if it's not the Recommended version. Delete the Recommended version of the service template if no other major versions or minor versions of the service template exist. A major version of a service template is a version that isn't backwards compatible. Delete a minor version of a service template if it's not the Recommended version. Delete a Recommended minor version of the service template if no other minor versions of the service template exist. A minor version of a service template is a version that's backwards compatible.

Arguments

major_version: The service template major version to delete.

minor_version: The service template minor version to delete.

template_name: The name of the service template.

Main.Proton.get_account_settings — Method

get_account_settings()
get_account_settings(params::Dict{String,<:Any})

Get detail data for the AWS Proton pipeline service role.

Main.Proton.get_environment — Method

get_environment(name)
get_environment(name, params::Dict{String,<:Any})

Get detail data for an environment.

Arguments

name: The name of the environment that you want to get the detail data for.

Main.Proton.get_environment_account_connection — Method

get_environment_account_connection(id)
get_environment_account_connection(id, params::Dict{String,<:Any})

In an environment account, view the detail data for an environment account connection. For more information, see Environment account connections in the AWS Proton Administration guide.

Arguments

id: The ID of the environment account connection.

Main.Proton.get_environment_template — Method

get_environment_template(name)
get_environment_template(name, params::Dict{String,<:Any})

Get detail data for an environment template.

Arguments

name: The name of the environment template that you want to get the detail data for.

Main.Proton.get_environment_template_version — Method

get_environment_template_version(major_version, minor_version, template_name)
get_environment_template_version(major_version, minor_version, template_name, params::Dict{String,<:Any})

View detail data for a major or minor version of an environment template.

Arguments

major_version: To view environment template major version detail data, include majorVersion.

minor_version: To view environment template minor version detail data, include minorVersion.

template_name: The name of the environment template.

Main.Proton.get_service — Method

get_service(name)
get_service(name, params::Dict{String,<:Any})

Get detail data for a service.

Arguments

name: The name of the service that you want to get the detail data for.

Main.Proton.get_service_instance — Method

get_service_instance(name, service_name)
get_service_instance(name, service_name, params::Dict{String,<:Any})

Get detail data for a service instance. A service instance is an instantiation of service template, which is running in a specific environment.

Arguments

name: The name of a service instance that you want to get the detail data for.

service_name: The name of the service that the service instance belongs to.

Main.Proton.get_service_template — Method

get_service_template(name)
get_service_template(name, params::Dict{String,<:Any})

Get detail data for a service template.

Arguments

name: The name of the service template that you want to get detail data for.

Main.Proton.get_service_template_version — Method

get_service_template_version(major_version, minor_version, template_name)
get_service_template_version(major_version, minor_version, template_name, params::Dict{String,<:Any})

View detail data for a major or minor version of a service template.

Arguments

major_version: To view service template major version detail data, include majorVersion.

minor_version: To view service template minor version detail data, include minorVersion.

template_name: The name of the service template.

Main.Proton.list_environment_account_connections — Method

list_environment_account_connections(requested_by)
list_environment_account_connections(requested_by, params::Dict{String,<:Any})

View a list of environment account connections. For more information, see Environment account connections in the AWS Proton Administration guide.

Arguments

requested_by: The type of account making the ListEnvironmentAccountConnections request.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"environmentName": The environment name that's associated with each listed environment account connection.

"maxResults": The maximum number of environment account connections to list.

"nextToken": A token to indicate the location of the next environment account connection in the array of environment account connections, after the list of environment account connections that was previously requested.

"statuses": The status details for each listed environment account connection.

Main.Proton.list_environment_template_versions — Method

list_environment_template_versions(template_name)
list_environment_template_versions(template_name, params::Dict{String,<:Any})

List major or minor versions of an environment template with detail data.

Arguments

template_name: The name of the environment template.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"majorVersion": To view a list of minor of versions under a major version of an environment template, include majorVersion. To view a list of major versions of an environment template, exclude majorVersion.

"maxResults": The maximum number of major or minor versions of an environment template to list.

"nextToken": A token to indicate the location of the next major or minor version in the array of major or minor versions of an environment template, after the list of major or minor versions that was previously requested.

Main.Proton.list_environment_templates — Method

list_environment_templates()
list_environment_templates(params::Dict{String,<:Any})

List environment templates.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of environment templates to list.

"nextToken": A token to indicate the location of the next environment template in the array of environment templates, after the list of environment templates that was previously requested.

Main.Proton.list_environments — Method

list_environments()
list_environments(params::Dict{String,<:Any})

List environments with detail data summaries.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"environmentTemplates": An array of the versions of the environment template.

"maxResults": The maximum number of environments to list.

"nextToken": A token to indicate the location of the next environment in the array of environments, after the list of environments that was previously requested.

Main.Proton.list_service_instances — Method

list_service_instances()
list_service_instances(params::Dict{String,<:Any})

List service instances with summaries of detail data.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of service instances to list.

"nextToken": A token to indicate the location of the next service in the array of service instances, after the list of service instances that was previously requested.

"serviceName": The name of the service that the service instance belongs to.

Main.Proton.list_service_template_versions — Method

list_service_template_versions(template_name)
list_service_template_versions(template_name, params::Dict{String,<:Any})

List major or minor versions of a service template with detail data.

Arguments

template_name: The name of the service template.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"majorVersion": To view a list of minor of versions under a major version of a service template, include majorVersion. To view a list of major versions of a service template, exclude majorVersion.

"maxResults": The maximum number of major or minor versions of a service template to list.

"nextToken": A token to indicate the location of the next major or minor version in the array of major or minor versions of a service template, after the list of major or minor versions that was previously requested.

Main.Proton.list_service_templates — Method

list_service_templates()
list_service_templates(params::Dict{String,<:Any})

List service templates with detail data.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of service templates to list.

"nextToken": A token to indicate the location of the next service template in the array of service templates, after the list of service templates previously requested.

Main.Proton.list_services — Method

list_services()
list_services(params::Dict{String,<:Any})

List services with summaries of detail data.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of services to list.

"nextToken": A token to indicate the location of the next service in the array of services, after the list of services that was previously requested.

Main.Proton.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

List tags for a resource. For more information, see AWS Proton resources and tagging in the AWS Proton Administration Guide or AWS Proton User Guide.

Arguments

resource_arn: The ARN of the resource for the listed tags.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of tags to list.

"nextToken": A token to indicate the location of the next resource tag in the array of resource tags, after the list of resource tags that was previously requested.

Main.Proton.reject_environment_account_connection — Method

reject_environment_account_connection(id)
reject_environment_account_connection(id, params::Dict{String,<:Any})

In a management account, reject an environment account connection from another environment account. After you reject an environment account connection request, you won’t be able to accept or use the rejected environment account connection. You can’t reject an environment account connection that is connected to an environment. For more information, see Environment account connections in the AWS Proton Administration guide.

Arguments

id: The ID of the environment account connection to reject.

Main.Proton.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Tag a resource. For more information, see AWS Proton resources and tagging in the AWS Proton Administration Guide or AWS Proton User Guide.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource that the resource tag is applied to.

tags: An array of resource tags to apply to a resource.

Main.Proton.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Remove a tag from a resource. For more information, see AWS Proton resources and tagging in the AWS Proton Administration Guide or AWS Proton User Guide.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource that the tag is to be removed from.

tag_keys: An array of tag keys indicating the resource tags to be removed from the resource.

Main.Proton.update_account_settings — Method

update_account_settings()
update_account_settings(params::Dict{String,<:Any})

Update the AWS Proton pipeline service account settings.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"pipelineServiceRoleArn": The Amazon Resource Name (ARN) of the AWS Proton pipeline service role.

Main.Proton.update_environment — Method

update_environment(deployment_type, name)
update_environment(deployment_type, name, params::Dict{String,<:Any})

Update an environment. If the environment is associated with an environment account connection, don't update or include the protonServiceRoleArn parameter to update or connect to an environment account connection. You can only update to a new environment account connection if it was created in the same environment account that the current environment account connection was created in and associated with the current environment. If the environment isn't associated with an environment account connection, don't update or include the environmentAccountConnectionId parameter to update or connect to an environment account connection. You can update either the environmentAccountConnectionId or protonServiceRoleArn parameter and value. You can’t update both. There are four modes for updating an environment as described in the following. The deploymentType field defines the mode. NONE In this mode, a deployment doesn't occur. Only the requested metadata parameters are updated. CURRENTVERSION In this mode, the environment is deployed and updated with the new spec that you provide. Only requested parameters are updated. Don’t include minor or major version parameters when you use this deployment-type. MINORVERSION In this mode, the environment is deployed and updated with the published, recommended (latest) minor version of the current major version in use, by default. You can also specify a different minor version of the current major version in use. MAJOR_VERSION In this mode, the environment is deployed and updated with the published, recommended (latest) major and minor version of the current template, by default. You can also specify a different major version that's higher than the major version in use and a minor version (optional).

Arguments

deployment_type: There are four modes for updating an environment as described in the following. The deploymentType field defines the mode. NONE In this mode, a deployment doesn't occur. Only the requested metadata parameters are updated. CURRENTVERSION In this mode, the environment is deployed and updated with the new spec that you provide. Only requested parameters are updated. Don’t include minor or major version parameters when you use this deployment-type. MINORVERSION In this mode, the environment is deployed and updated with the published, recommended (latest) minor version of the current major version in use, by default. You can also specify a different minor version of the current major version in use. MAJOR_VERSION In this mode, the environment is deployed and updated with the published, recommended (latest) major and minor version of the current template, by default. You can also specify a different major version that is higher than the major version in use and a minor version (optional).

name: The name of the environment to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A description of the environment update.

"environmentAccountConnectionId": The ID of the environment account connection. You can only update to a new environment account connection if it was created in the same environment account that the current environment account connection was created in and associated with the current environment.

"protonServiceRoleArn": The ARN of the AWS Proton service role that allows AWS Proton to make API calls to other services your behalf.

"spec": The formatted specification that defines the update.

"templateMajorVersion": The ID of the major version of the environment to update.

"templateMinorVersion": The ID of the minor version of the environment to update.

Main.Proton.update_environment_account_connection — Method

update_environment_account_connection(id, role_arn)
update_environment_account_connection(id, role_arn, params::Dict{String,<:Any})

In an environment account, update an environment account connection to use a new IAM role. For more information, see Environment account connections in the AWS Proton Administration guide.

Arguments

id: The ID of the environment account connection to update.

role_arn: The Amazon Resource Name (ARN) of the IAM service role that is associated with the environment account connection to update.

Main.Proton.update_environment_template — Method

update_environment_template(name)
update_environment_template(name, params::Dict{String,<:Any})

Update an environment template.

Arguments

name: The name of the environment template to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A description of the environment template update.

"displayName": The name of the environment template to update as displayed in the developer interface.

Main.Proton.update_environment_template_version — Method

update_environment_template_version(major_version, minor_version, template_name)
update_environment_template_version(major_version, minor_version, template_name, params::Dict{String,<:Any})

Update a major or minor version of an environment template.

Arguments

major_version: To update a major version of an environment template, include majorVersion.

minor_version: To update a minor version of an environment template, include minorVersion.

template_name: The name of the environment template.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A description of environment template version to update.

"status": The status of the environment template minor version to update.

Main.Proton.update_service — Method

update_service(name)
update_service(name, params::Dict{String,<:Any})

Edit a service description or use a spec to add and delete service instances. Existing service instances and the service pipeline can't be edited using this API. They can only be deleted. Use the description parameter to modify the description. Edit the spec parameter to add or delete instances.

Arguments

name: The name of the service to edit.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The edited service description.

"spec": Lists the service instances to add and the existing service instances to remain. Omit the existing service instances to delete from the list. Don't include edits to the existing service instances or pipeline. For more information, see Edit a service in the AWS Proton Administration Guide or the AWS Proton User Guide.

Main.Proton.update_service_instance — Method

update_service_instance(deployment_type, name, service_name)
update_service_instance(deployment_type, name, service_name, params::Dict{String,<:Any})

Update a service instance. There are four modes for updating a service instance as described in the following. The deploymentType field defines the mode. NONE In this mode, a deployment doesn't occur. Only the requested metadata parameters are updated. CURRENTVERSION In this mode, the service instance is deployed and updated with the new spec that you provide. Only requested parameters are updated. Don’t include minor or major version parameters when you use this deployment-type. MINORVERSION In this mode, the service instance is deployed and updated with the published, recommended (latest) minor version of the current major version in use, by default. You can also specify a different minor version of the current major version in use. MAJOR_VERSION In this mode, the service instance is deployed and updated with the published, recommended (latest) major and minor version of the current template, by default. You can also specify a different major version that is higher than the major version in use and a minor version (optional).

Arguments

deployment_type: The deployment type. There are four modes for updating a service instance as described in the following. The deploymentType field defines the mode. NONE In this mode, a deployment doesn't occur. Only the requested metadata parameters are updated. CURRENTVERSION In this mode, the service instance is deployed and updated with the new spec that you provide. Only requested parameters are updated. Don’t include minor or major version parameters when you use this deployment-type. MINORVERSION In this mode, the service instance is deployed and updated with the published, recommended (latest) minor version of the current major version in use, by default. You can also specify a different minor version of the current major version in use. MAJOR_VERSION In this mode, the service instance is deployed and updated with the published, recommended (latest) major and minor version of the current template, by default. You can also specify a different major version that is higher than the major version in use and a minor version (optional).

name: The name of the service instance to update.

service_name: The name of the service that the service instance belongs to.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"spec": The formatted specification that defines the service instance update.

"templateMajorVersion": The major version of the service template to update.

"templateMinorVersion": The minor version of the service template to update.

Main.Proton.update_service_pipeline — Method

update_service_pipeline(deployment_type, service_name, spec)
update_service_pipeline(deployment_type, service_name, spec, params::Dict{String,<:Any})

Update the service pipeline. There are four modes for updating a service pipeline as described in the following. The deploymentType field defines the mode. NONE In this mode, a deployment doesn't occur. Only the requested metadata parameters are updated. CURRENTVERSION In this mode, the service pipeline is deployed and updated with the new spec that you provide. Only requested parameters are updated. Don’t include minor or major version parameters when you use this deployment-type. MINORVERSION In this mode, the service pipeline is deployed and updated with the published, recommended (latest) minor version of the current major version in use, by default. You can also specify a different minor version of the current major version in use. MAJOR_VERSION In this mode, the service pipeline is deployed and updated with the published, recommended (latest) major and minor version of the current template by default. You can also specify a different major version that is higher than the major version in use and a minor version (optional).

Arguments

deployment_type: The deployment type. There are four modes for updating a service pipeline as described in the following. The deploymentType field defines the mode. NONE In this mode, a deployment doesn't occur. Only the requested metadata parameters are updated. CURRENTVERSION In this mode, the service pipeline is deployed and updated with the new spec that you provide. Only requested parameters are updated. Don’t include minor or major version parameters when you use this deployment-type. MINORVERSION In this mode, the service pipeline is deployed and updated with the published, recommended (latest) minor version of the current major version in use, by default. You can also specify a different minor version of the current major version in use. MAJOR_VERSION In this mode, the service pipeline is deployed and updated with the published, recommended (latest) major and minor version of the current template, by default. You can also specify a different major version that is higher than the major version in use and a minor version (optional).

service_name: The name of the service to that the pipeline is associated with.

spec: The spec for the service pipeline to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"templateMajorVersion": The major version of the service template that was used to create the service that the pipeline is associated with.

"templateMinorVersion": The minor version of the service template that was used to create the service that the pipeline is associated with.

Main.Proton.update_service_template — Method

update_service_template(name)
update_service_template(name, params::Dict{String,<:Any})

Update a service template.

Arguments

name: The name of the service template to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A description of the service template update.

"displayName": The name of the service template to update as displayed in the developer interface.

Main.Proton.update_service_template_version — Method

update_service_template_version(major_version, minor_version, template_name)
update_service_template_version(major_version, minor_version, template_name, params::Dict{String,<:Any})

Update a major or minor version of a service template.

Arguments

major_version: To update a major version of a service template, include majorVersion.

minor_version: To update a minor version of a service template, include minorVersion.

template_name: The name of the service template.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"compatibleEnvironmentTemplates": An array of compatible environment names for a service template major or minor version to update.

"description": A description of a service template version to update.

"status": The status of the service template minor version to update.