Qconnect

This page documents function available when using the Qconnect module, created with @service Qconnect.

Index

Main.Qconnect.create_assistant

Main.Qconnect.create_assistant_association

Main.Qconnect.create_content

Main.Qconnect.create_knowledge_base

Main.Qconnect.create_quick_response

Main.Qconnect.create_session

Main.Qconnect.delete_assistant

Main.Qconnect.delete_assistant_association

Main.Qconnect.delete_content

Main.Qconnect.delete_import_job

Main.Qconnect.delete_knowledge_base

Main.Qconnect.delete_quick_response

Main.Qconnect.get_assistant

Main.Qconnect.get_assistant_association

Main.Qconnect.get_content

Main.Qconnect.get_content_summary

Main.Qconnect.get_import_job

Main.Qconnect.get_knowledge_base

Main.Qconnect.get_quick_response

Main.Qconnect.get_recommendations

Main.Qconnect.get_session

Main.Qconnect.list_assistant_associations

Main.Qconnect.list_assistants

Main.Qconnect.list_contents

Main.Qconnect.list_import_jobs

Main.Qconnect.list_knowledge_bases

Main.Qconnect.list_quick_responses

Main.Qconnect.list_tags_for_resource

Main.Qconnect.notify_recommendations_received

Main.Qconnect.put_feedback

Main.Qconnect.query_assistant

Main.Qconnect.remove_knowledge_base_template_uri

Main.Qconnect.search_content

Main.Qconnect.search_quick_responses

Main.Qconnect.search_sessions

Main.Qconnect.start_content_upload

Main.Qconnect.start_import_job

Main.Qconnect.tag_resource

Main.Qconnect.untag_resource

Main.Qconnect.update_content

Main.Qconnect.update_knowledge_base_template_uri

Main.Qconnect.update_quick_response

Main.Qconnect.update_session

Documentation

Main.Qconnect.create_assistant — Method

create_assistant(name, type)
create_assistant(name, type, params::Dict{String,<:Any})

Creates an Amazon Q in Connect assistant.

Arguments

name: The name of the assistant.

type: The type of assistant.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If not provided, the Amazon Web Services SDK populates this field. For more information about idempotency, see Making retries safe with idempotent APIs.

"description": The description of the assistant.

"serverSideEncryptionConfiguration": The configuration information for the customer managed key used for encryption. The customer managed key must have a policy that allows kms:CreateGrant, kms:DescribeKey, kms:Decrypt, and kms:GenerateDataKey* permissions to the IAM identity using the key to invoke Amazon Q in Connect. To use Amazon Q in Connect with chat, the key policy must also allow kms:Decrypt, kms:GenerateDataKey*, and kms:DescribeKey permissions to the connect.amazonaws.com service principal. For more information about setting up a customer managed key for Amazon Q in Connect, see Enable Amazon Q in Connect for your instance.

"tags": The tags used to organize, track, or control access for this resource.

Main.Qconnect.create_assistant_association — Method

create_assistant_association(assistant_id, association, association_type)
create_assistant_association(assistant_id, association, association_type, params::Dict{String,<:Any})

Creates an association between an Amazon Q in Connect assistant and another resource. Currently, the only supported association is with a knowledge base. An assistant can have only a single association.

Arguments

assistant_id: The identifier of the Amazon Q in Connect assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

association: The identifier of the associated resource.

association_type: The type of association.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If not provided, the Amazon Web Services SDK populates this field. For more information about idempotency, see Making retries safe with idempotent APIs.

"tags": The tags used to organize, track, or control access for this resource.

Main.Qconnect.create_content — Method

create_content(knowledge_base_id, name, upload_id)
create_content(knowledge_base_id, name, upload_id, params::Dict{String,<:Any})

Creates Amazon Q in Connect content. Before to calling this API, use StartContentUpload to upload an asset.

Arguments

knowledge_base_id: The identifier of the knowledge base. This should not be a QUICK_RESPONSES type knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

name: The name of the content. Each piece of content in a knowledge base must have a unique name. You can retrieve a piece of content using only its knowledge base and its name with the SearchContent API.

upload_id: A pointer to the uploaded asset. This value is returned by StartContentUpload.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If not provided, the Amazon Web Services SDK populates this field. For more information about idempotency, see Making retries safe with idempotent APIs.

"metadata": A key/value map to store attributes without affecting tagging or recommendations. For example, when synchronizing data between an external system and Amazon Q in Connect, you can store an external version identifier as metadata to utilize for determining drift.

"overrideLinkOutUri": The URI you want to use for the article. If the knowledge base has a templateUri, setting this argument overrides it for this piece of content.

"tags": The tags used to organize, track, or control access for this resource.

"title": The title of the content. If not set, the title is equal to the name.

Main.Qconnect.create_knowledge_base — Method

create_knowledge_base(knowledge_base_type, name)
create_knowledge_base(knowledge_base_type, name, params::Dict{String,<:Any})

Creates a knowledge base. When using this API, you cannot reuse Amazon AppIntegrations DataIntegrations with external knowledge bases such as Salesforce and ServiceNow. If you do, you'll get an InvalidRequestException error. For example, you're programmatically managing your external knowledge base, and you want to add or remove one of the fields that is being ingested from Salesforce. Do the following: Call DeleteKnowledgeBase. Call DeleteDataIntegration. Call CreateDataIntegration to recreate the DataIntegration or a create different one. Call CreateKnowledgeBase.

Arguments

knowledge_base_type: The type of knowledge base. Only CUSTOM knowledge bases allow you to upload your own content. EXTERNAL knowledge bases support integrations with third-party systems whose content is synchronized automatically.

name: The name of the knowledge base.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If not provided, the Amazon Web Services SDK populates this field. For more information about idempotency, see Making retries safe with idempotent APIs.

"description": The description.

"renderingConfiguration": Information about how to render the content.

"serverSideEncryptionConfiguration": The configuration information for the customer managed key used for encryption. This KMS key must have a policy that allows kms:CreateGrant, kms:DescribeKey, kms:Decrypt, and kms:GenerateDataKey* permissions to the IAM identity using the key to invoke Amazon Q in Connect. For more information about setting up a customer managed key for Amazon Q in Connect, see Enable Amazon Q in Connect for your instance.

"sourceConfiguration": The source of the knowledge base content. Only set this argument for EXTERNAL knowledge bases.

"tags": The tags used to organize, track, or control access for this resource.

Main.Qconnect.create_quick_response — Method

create_quick_response(content, knowledge_base_id, name)
create_quick_response(content, knowledge_base_id, name, params::Dict{String,<:Any})

Creates an Amazon Q in Connect quick response.

Arguments

content: The content of the quick response.

knowledge_base_id: The identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

name: The name of the quick response.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"channels": The Amazon Connect channels this quick response applies to.

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If not provided, the Amazon Web Services SDK populates this field. For more information about idempotency, see Making retries safe with idempotent APIs.

"contentType": The media type of the quick response content. Use application/x.quickresponse;format=plain for a quick response written in plain text. Use application/x.quickresponse;format=markdown for a quick response written in richtext.

"description": The description of the quick response.

"groupingConfiguration": The configuration information of the user groups that the quick response is accessible to.

"isActive": Whether the quick response is active.

"language": The language code value for the language in which the quick response is written. The supported language codes include deDE, enUS, esES, frFR, idID, itIT, jaJP, koKR, ptBR, zhCN, zh_TW

"shortcutKey": The shortcut key of the quick response. The value should be unique across the knowledge base.

"tags": The tags used to organize, track, or control access for this resource.

Main.Qconnect.create_session — Method

create_session(assistant_id, name)
create_session(assistant_id, name, params::Dict{String,<:Any})

Creates a session. A session is a contextual container used for generating recommendations. Amazon Connect creates a new Amazon Q in Connect session for each contact on which Amazon Q in Connect is enabled.

Arguments

assistant_id: The identifier of the Amazon Q in Connect assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

name: The name of the session.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If not provided, the Amazon Web Services SDK populates this field. For more information about idempotency, see Making retries safe with idempotent APIs.

"description": The description.

"tagFilter": An object that can be used to specify Tag conditions.

"tags": The tags used to organize, track, or control access for this resource.

Main.Qconnect.delete_assistant — Method

delete_assistant(assistant_id)
delete_assistant(assistant_id, params::Dict{String,<:Any})

Deletes an assistant.

Arguments

assistant_id: The identifier of the Amazon Q in Connect assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

Main.Qconnect.delete_assistant_association — Method

delete_assistant_association(assistant_association_id, assistant_id)
delete_assistant_association(assistant_association_id, assistant_id, params::Dict{String,<:Any})

Deletes an assistant association.

Arguments

assistant_association_id: The identifier of the assistant association. Can be either the ID or the ARN. URLs cannot contain the ARN.

assistant_id: The identifier of the Amazon Q in Connect assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

Main.Qconnect.delete_content — Method

delete_content(content_id, knowledge_base_id)
delete_content(content_id, knowledge_base_id, params::Dict{String,<:Any})

Deletes the content.

Arguments

content_id: The identifier of the content. Can be either the ID or the ARN. URLs cannot contain the ARN.

knowledge_base_id: The identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

Main.Qconnect.delete_import_job — Method

delete_import_job(import_job_id, knowledge_base_id)
delete_import_job(import_job_id, knowledge_base_id, params::Dict{String,<:Any})

Deletes the quick response import job.

Arguments

import_job_id: The identifier of the import job to be deleted.

knowledge_base_id: The identifier of the knowledge base.

Main.Qconnect.delete_knowledge_base — Method

delete_knowledge_base(knowledge_base_id)
delete_knowledge_base(knowledge_base_id, params::Dict{String,<:Any})

Deletes the knowledge base. When you use this API to delete an external knowledge base such as Salesforce or ServiceNow, you must also delete the Amazon AppIntegrations DataIntegration. This is because you can't reuse the DataIntegration after it's been associated with an external knowledge base. However, you can delete and recreate it. See DeleteDataIntegration and CreateDataIntegration in the Amazon AppIntegrations API Reference.

Arguments

knowledge_base_id: The knowledge base to delete content from. Can be either the ID or the ARN. URLs cannot contain the ARN.

Main.Qconnect.delete_quick_response — Method

delete_quick_response(knowledge_base_id, quick_response_id)
delete_quick_response(knowledge_base_id, quick_response_id, params::Dict{String,<:Any})

Deletes a quick response.

Arguments

knowledge_base_id: The knowledge base from which the quick response is deleted. The identifier of the knowledge base.

quick_response_id: The identifier of the quick response to delete.

Main.Qconnect.get_assistant — Method

get_assistant(assistant_id)
get_assistant(assistant_id, params::Dict{String,<:Any})

Retrieves information about an assistant.

Arguments

assistant_id: The identifier of the Amazon Q in Connect assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

Main.Qconnect.get_assistant_association — Method

get_assistant_association(assistant_association_id, assistant_id)
get_assistant_association(assistant_association_id, assistant_id, params::Dict{String,<:Any})

Retrieves information about an assistant association.

Arguments

assistant_association_id: The identifier of the assistant association. Can be either the ID or the ARN. URLs cannot contain the ARN.

assistant_id: The identifier of the Amazon Q in Connect assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

Main.Qconnect.get_content — Method

get_content(content_id, knowledge_base_id)
get_content(content_id, knowledge_base_id, params::Dict{String,<:Any})

Retrieves content, including a pre-signed URL to download the content.

Arguments

content_id: The identifier of the content. Can be either the ID or the ARN. URLs cannot contain the ARN.

knowledge_base_id: The identifier of the knowledge base. This should not be a QUICK_RESPONSES type knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

Main.Qconnect.get_content_summary — Method

get_content_summary(content_id, knowledge_base_id)
get_content_summary(content_id, knowledge_base_id, params::Dict{String,<:Any})

Retrieves summary information about the content.

Arguments

content_id: The identifier of the content. Can be either the ID or the ARN. URLs cannot contain the ARN.

knowledge_base_id: The identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

Main.Qconnect.get_import_job — Method

get_import_job(import_job_id, knowledge_base_id)
get_import_job(import_job_id, knowledge_base_id, params::Dict{String,<:Any})

Retrieves the started import job.

Arguments

import_job_id: The identifier of the import job to retrieve.

knowledge_base_id: The identifier of the knowledge base that the import job belongs to.

Main.Qconnect.get_knowledge_base — Method

get_knowledge_base(knowledge_base_id)
get_knowledge_base(knowledge_base_id, params::Dict{String,<:Any})

Retrieves information about the knowledge base.

Arguments

knowledge_base_id: The identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

Main.Qconnect.get_quick_response — Method

get_quick_response(knowledge_base_id, quick_response_id)
get_quick_response(knowledge_base_id, quick_response_id, params::Dict{String,<:Any})

Retrieves the quick response.

Arguments

knowledge_base_id: The identifier of the knowledge base. This should be a QUICK_RESPONSES type knowledge base.

quick_response_id: The identifier of the quick response.

Main.Qconnect.get_recommendations — Method

get_recommendations(assistant_id, session_id)
get_recommendations(assistant_id, session_id, params::Dict{String,<:Any})

This API will be discontinued starting June 1, 2024. To receive generative responses after March 1, 2024, you will need to create a new Assistant in the Amazon Connect console and integrate the Amazon Q in Connect JavaScript library (amazon-q-connectjs) into your applications. Retrieves recommendations for the specified session. To avoid retrieving the same recommendations in subsequent calls, use NotifyRecommendationsReceived. This API supports long-polling behavior with the waitTimeSeconds parameter. Short poll is the default behavior and only returns recommendations already available. To perform a manual query against an assistant, use QueryAssistant.

Arguments

assistant_id: The identifier of the Amazon Q in Connect assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

session_id: The identifier of the session. Can be either the ID or the ARN. URLs cannot contain the ARN.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return per page.

"waitTimeSeconds": The duration (in seconds) for which the call waits for a recommendation to be made available before returning. If a recommendation is available, the call returns sooner than WaitTimeSeconds. If no messages are available and the wait time expires, the call returns successfully with an empty list.

Main.Qconnect.get_session — Method

get_session(assistant_id, session_id)
get_session(assistant_id, session_id, params::Dict{String,<:Any})

Retrieves information for a specified session.

Arguments

assistant_id: The identifier of the Amazon Q in Connect assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

session_id: The identifier of the session. Can be either the ID or the ARN. URLs cannot contain the ARN.

Main.Qconnect.list_assistant_associations — Method

list_assistant_associations(assistant_id)
list_assistant_associations(assistant_id, params::Dict{String,<:Any})

Lists information about assistant associations.

Arguments

assistant_id: The identifier of the Amazon Q in Connect assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return per page.

"nextToken": The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

Main.Qconnect.list_assistants — Method

list_assistants()
list_assistants(params::Dict{String,<:Any})

Lists information about assistants.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return per page.

"nextToken": The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

Main.Qconnect.list_contents — Method

list_contents(knowledge_base_id)
list_contents(knowledge_base_id, params::Dict{String,<:Any})

Lists the content.

Arguments

knowledge_base_id: The identifier of the knowledge base. This should not be a QUICK_RESPONSES type knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return per page.

"nextToken": The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

Main.Qconnect.list_import_jobs — Method

list_import_jobs(knowledge_base_id)
list_import_jobs(knowledge_base_id, params::Dict{String,<:Any})

Lists information about import jobs.

Arguments

knowledge_base_id: The identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return per page.

"nextToken": The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

Main.Qconnect.list_knowledge_bases — Method

list_knowledge_bases()
list_knowledge_bases(params::Dict{String,<:Any})

Lists the knowledge bases.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return per page.

"nextToken": The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

Main.Qconnect.list_quick_responses — Method

list_quick_responses(knowledge_base_id)
list_quick_responses(knowledge_base_id, params::Dict{String,<:Any})

Lists information about quick response.

Arguments

knowledge_base_id: The identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return per page.

"nextToken": The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

Main.Qconnect.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists the tags for the specified resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

Main.Qconnect.notify_recommendations_received — Method

notify_recommendations_received(assistant_id, recommendation_ids, session_id)
notify_recommendations_received(assistant_id, recommendation_ids, session_id, params::Dict{String,<:Any})

Removes the specified recommendations from the specified assistant's queue of newly available recommendations. You can use this API in conjunction with GetRecommendations and a waitTimeSeconds input for long-polling behavior and avoiding duplicate recommendations.

Arguments

assistant_id: The identifier of the Amazon Q in Connect assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

recommendation_ids: The identifiers of the recommendations.

session_id: The identifier of the session. Can be either the ID or the ARN. URLs cannot contain the ARN.

Main.Qconnect.put_feedback — Method

put_feedback(assistant_id, content_feedback, target_id, target_type)
put_feedback(assistant_id, content_feedback, target_id, target_type, params::Dict{String,<:Any})

Provides feedback against the specified assistant for the specified target. This API only supports generative targets.

Arguments

assistant_id: The identifier of the Amazon Q in Connect assistant.

content_feedback: Information about the feedback provided.

target_id: The identifier of the feedback target.

target_type: The type of the feedback target.

Main.Qconnect.query_assistant — Method

query_assistant(assistant_id, query_text)
query_assistant(assistant_id, query_text, params::Dict{String,<:Any})

This API will be discontinued starting June 1, 2024. To receive generative responses after March 1, 2024, you will need to create a new Assistant in the Amazon Connect console and integrate the Amazon Q in Connect JavaScript library (amazon-q-connectjs) into your applications. Performs a manual search against the specified assistant. To retrieve recommendations for an assistant, use GetRecommendations.

Arguments

assistant_id: The identifier of the Amazon Q in Connect assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

query_text: The text to search for.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return per page.

"nextToken": The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

"queryCondition": Information about how to query content.

"sessionId": The identifier of the Amazon Q in Connect session. Can be either the ID or the ARN. URLs cannot contain the ARN.

Main.Qconnect.remove_knowledge_base_template_uri — Method

remove_knowledge_base_template_uri(knowledge_base_id)
remove_knowledge_base_template_uri(knowledge_base_id, params::Dict{String,<:Any})

Removes a URI template from a knowledge base.

Arguments

knowledge_base_id: The identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

Main.Qconnect.search_content — Method

search_content(knowledge_base_id, search_expression)
search_content(knowledge_base_id, search_expression, params::Dict{String,<:Any})

Searches for content in a specified knowledge base. Can be used to get a specific content resource by its name.

Arguments

knowledge_base_id: The identifier of the knowledge base. This should not be a QUICK_RESPONSES type knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

search_expression: The search expression to filter results.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return per page.

"nextToken": The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

Main.Qconnect.search_quick_responses — Method

search_quick_responses(knowledge_base_id, search_expression)
search_quick_responses(knowledge_base_id, search_expression, params::Dict{String,<:Any})

Searches existing Amazon Q in Connect quick responses in an Amazon Q in Connect knowledge base.

Arguments

knowledge_base_id: The identifier of the knowledge base. This should be a QUICK_RESPONSES type knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

search_expression: The search expression for querying the quick response.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"attributes": The user-defined Amazon Connect contact attributes to be resolved when search results are returned.

"maxResults": The maximum number of results to return per page.

"nextToken": The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

Main.Qconnect.search_sessions — Method

search_sessions(assistant_id, search_expression)
search_sessions(assistant_id, search_expression, params::Dict{String,<:Any})

Searches for sessions.

Arguments

assistant_id: The identifier of the Amazon Q in Connect assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

search_expression: The search expression to filter results.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return per page.

"nextToken": The token for the next set of results. Use the value returned in the previous response in the next request to retrieve the next set of results.

Main.Qconnect.start_content_upload — Method

start_content_upload(content_type, knowledge_base_id)
start_content_upload(content_type, knowledge_base_id, params::Dict{String,<:Any})

Get a URL to upload content to a knowledge base. To upload content, first make a PUT request to the returned URL with your file, making sure to include the required headers. Then use CreateContent to finalize the content creation process or UpdateContent to modify an existing resource. You can only upload content to a knowledge base of type CUSTOM.

Arguments

content_type: The type of content to upload.

knowledge_base_id: The identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"presignedUrlTimeToLive": The expected expiration time of the generated presigned URL, specified in minutes.

Main.Qconnect.start_import_job — Method

start_import_job(import_job_type, knowledge_base_id, upload_id)
start_import_job(import_job_type, knowledge_base_id, upload_id, params::Dict{String,<:Any})

Start an asynchronous job to import Amazon Q in Connect resources from an uploaded source file. Before calling this API, use StartContentUpload to upload an asset that contains the resource data. For importing Amazon Q in Connect quick responses, you need to upload a csv file including the quick responses. For information about how to format the csv file for importing quick responses, see Import quick responses.

Arguments

import_job_type: The type of the import job. For importing quick response resource, set the value to QUICK_RESPONSES.

knowledge_base_id: The identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN. For importing Amazon Q in Connect quick responses, this should be a QUICK_RESPONSES type knowledge base.

upload_id: A pointer to the uploaded asset. This value is returned by StartContentUpload.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": The tags used to organize, track, or control access for this resource.

"externalSourceConfiguration": The configuration information of the external source that the resource data are imported from.

"metadata": The metadata fields of the imported Amazon Q in Connect resources.

Main.Qconnect.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Adds the specified tags to the specified resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

tags: The tags used to organize, track, or control access for this resource.

Main.Qconnect.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes the specified tags from the specified resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

tag_keys: The tag keys.

Main.Qconnect.update_content — Method

update_content(content_id, knowledge_base_id)
update_content(content_id, knowledge_base_id, params::Dict{String,<:Any})

Updates information about the content.

Arguments

content_id: The identifier of the content. Can be either the ID or the ARN. URLs cannot contain the ARN.

knowledge_base_id: The identifier of the knowledge base. This should not be a QUICK_RESPONSES type knowledge base. Can be either the ID or the ARN

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"metadata": A key/value map to store attributes without affecting tagging or recommendations. For example, when synchronizing data between an external system and Amazon Q in Connect, you can store an external version identifier as metadata to utilize for determining drift.

"overrideLinkOutUri": The URI for the article. If the knowledge base has a templateUri, setting this argument overrides it for this piece of content. To remove an existing overrideLinkOurUri, exclude this argument and set removeOverrideLinkOutUri to true.

"removeOverrideLinkOutUri": Unset the existing overrideLinkOutUri if it exists.

"revisionId": The revisionId of the content resource to update, taken from an earlier call to GetContent, GetContentSummary, SearchContent, or ListContents. If included, this argument acts as an optimistic lock to ensure content was not modified since it was last read. If it has been modified, this API throws a PreconditionFailedException.

"title": The title of the content.

"uploadId": A pointer to the uploaded asset. This value is returned by StartContentUpload.

Main.Qconnect.update_knowledge_base_template_uri — Method

update_knowledge_base_template_uri(knowledge_base_id, template_uri)
update_knowledge_base_template_uri(knowledge_base_id, template_uri, params::Dict{String,<:Any})

Updates the template URI of a knowledge base. This is only supported for knowledge bases of type EXTERNAL. Include a single variable in {variable} format; this interpolated by Amazon Q in Connect using ingested content. For example, if you ingest a Salesforce article, it has an Id value, and you can set the template URI to https://myInstanceName.lightning.force.com/lightning/r/Knowledge__kav/{Id}/view.

Arguments

knowledge_base_id: The identifier of the knowledge base. This should not be a QUICK_RESPONSES type knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

template_uri: The template URI to update.

Main.Qconnect.update_quick_response — Method

update_quick_response(knowledge_base_id, quick_response_id)
update_quick_response(knowledge_base_id, quick_response_id, params::Dict{String,<:Any})

Updates an existing Amazon Q in Connect quick response.

Arguments

knowledge_base_id: The identifier of the knowledge base. Can be either the ID or the ARN. URLs cannot contain the ARN.

quick_response_id: The identifier of the quick response.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"channels": The Amazon Connect contact channels this quick response applies to. The supported contact channel types include Chat.

"content": The updated content of the quick response.

"contentType": The media type of the quick response content. Use application/x.quickresponse;format=plain for quick response written in plain text. Use application/x.quickresponse;format=markdown for quick response written in richtext.

"description": The updated description of the quick response.

"groupingConfiguration": The updated grouping configuration of the quick response.

"isActive": Whether the quick response is active.

"language": The language code value for the language in which the quick response is written. The supported language codes include deDE, enUS, esES, frFR, idID, itIT, jaJP, koKR, ptBR, zhCN, zh_TW

"name": The name of the quick response.

"removeDescription": Whether to remove the description from the quick response.

"removeGroupingConfiguration": Whether to remove the grouping configuration of the quick response.

"removeShortcutKey": Whether to remove the shortcut key of the quick response.

"shortcutKey": The shortcut key of the quick response. The value should be unique across the knowledge base.

Main.Qconnect.update_session — Method

update_session(assistant_id, session_id)
update_session(assistant_id, session_id, params::Dict{String,<:Any})

Updates a session. A session is a contextual container used for generating recommendations. Amazon Connect updates the existing Amazon Q in Connect session for each contact on which Amazon Q in Connect is enabled.

Arguments

assistant_id: The identifier of the Amazon Q in Connect assistant. Can be either the ID or the ARN. URLs cannot contain the ARN.

session_id: The identifier of the session. Can be either the ID or the ARN. URLs cannot contain the ARN.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The description.

"tagFilter": An object that can be used to specify Tag conditions.