Robomaker

This page documents function available when using the Robomaker module, created with @service Robomaker.

Index

Main.Robomaker.batch_delete_worlds

Main.Robomaker.batch_describe_simulation_job

Main.Robomaker.cancel_deployment_job

Main.Robomaker.cancel_simulation_job

Main.Robomaker.cancel_simulation_job_batch

Main.Robomaker.cancel_world_export_job

Main.Robomaker.cancel_world_generation_job

Main.Robomaker.create_deployment_job

Main.Robomaker.create_fleet

Main.Robomaker.create_robot

Main.Robomaker.create_robot_application

Main.Robomaker.create_robot_application_version

Main.Robomaker.create_simulation_application

Main.Robomaker.create_simulation_application_version

Main.Robomaker.create_simulation_job

Main.Robomaker.create_world_export_job

Main.Robomaker.create_world_generation_job

Main.Robomaker.create_world_template

Main.Robomaker.delete_fleet

Main.Robomaker.delete_robot

Main.Robomaker.delete_robot_application

Main.Robomaker.delete_simulation_application

Main.Robomaker.delete_world_template

Main.Robomaker.deregister_robot

Main.Robomaker.describe_deployment_job

Main.Robomaker.describe_fleet

Main.Robomaker.describe_robot

Main.Robomaker.describe_robot_application

Main.Robomaker.describe_simulation_application

Main.Robomaker.describe_simulation_job

Main.Robomaker.describe_simulation_job_batch

Main.Robomaker.describe_world

Main.Robomaker.describe_world_export_job

Main.Robomaker.describe_world_generation_job

Main.Robomaker.describe_world_template

Main.Robomaker.get_world_template_body

Main.Robomaker.list_deployment_jobs

Main.Robomaker.list_fleets

Main.Robomaker.list_robot_applications

Main.Robomaker.list_robots

Main.Robomaker.list_simulation_applications

Main.Robomaker.list_simulation_job_batches

Main.Robomaker.list_simulation_jobs

Main.Robomaker.list_tags_for_resource

Main.Robomaker.list_world_export_jobs

Main.Robomaker.list_world_generation_jobs

Main.Robomaker.list_world_templates

Main.Robomaker.list_worlds

Main.Robomaker.register_robot

Main.Robomaker.restart_simulation_job

Main.Robomaker.start_simulation_job_batch

Main.Robomaker.sync_deployment_job

Main.Robomaker.tag_resource

Main.Robomaker.untag_resource

Main.Robomaker.update_robot_application

Main.Robomaker.update_simulation_application

Main.Robomaker.update_world_template

Documentation

Main.Robomaker.batch_delete_worlds — Method

batch_delete_worlds(worlds)
batch_delete_worlds(worlds, params::Dict{String,<:Any})

Deletes one or more worlds in a batch operation.

Arguments

worlds: A list of Amazon Resource Names (arns) that correspond to worlds to delete.

Main.Robomaker.batch_describe_simulation_job — Method

batch_describe_simulation_job(jobs)
batch_describe_simulation_job(jobs, params::Dict{String,<:Any})

Describes one or more simulation jobs.

Arguments

jobs: A list of Amazon Resource Names (ARNs) of simulation jobs to describe.

Main.Robomaker.cancel_deployment_job — Method

cancel_deployment_job(job)
cancel_deployment_job(job, params::Dict{String,<:Any})

Cancels the specified deployment job.

Arguments

job: The deployment job ARN to cancel.

Main.Robomaker.cancel_simulation_job — Method

cancel_simulation_job(job)
cancel_simulation_job(job, params::Dict{String,<:Any})

Cancels the specified simulation job.

Arguments

job: The simulation job ARN to cancel.

Main.Robomaker.cancel_simulation_job_batch — Method

cancel_simulation_job_batch(batch)
cancel_simulation_job_batch(batch, params::Dict{String,<:Any})

Cancels a simulation job batch. When you cancel a simulation job batch, you are also cancelling all of the active simulation jobs created as part of the batch.

Arguments

batch: The id of the batch to cancel.

Main.Robomaker.cancel_world_export_job — Method

cancel_world_export_job(job)
cancel_world_export_job(job, params::Dict{String,<:Any})

Cancels the specified export job.

Arguments

job: The Amazon Resource Name (arn) of the world export job to cancel.

Main.Robomaker.cancel_world_generation_job — Method

cancel_world_generation_job(job)
cancel_world_generation_job(job, params::Dict{String,<:Any})

Cancels the specified world generator job.

Arguments

job: The Amazon Resource Name (arn) of the world generator job to cancel.

Main.Robomaker.create_deployment_job — Method

create_deployment_job(client_request_token, deployment_application_configs, fleet)
create_deployment_job(client_request_token, deployment_application_configs, fleet, params::Dict{String,<:Any})

Deploys a specific version of a robot application to robots in a fleet. The robot application must have a numbered applicationVersion for consistency reasons. To create a new version, use CreateRobotApplicationVersion or see Creating a Robot Application Version. After 90 days, deployment jobs expire and will be deleted. They will no longer be accessible.

Arguments

client_request_token: Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

deployment_application_configs: The deployment application configuration.

fleet: The Amazon Resource Name (ARN) of the fleet to deploy.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"deploymentConfig": The requested deployment configuration.

"tags": A map that contains tag keys and tag values that are attached to the deployment job.

Main.Robomaker.create_fleet — Method

create_fleet(name)
create_fleet(name, params::Dict{String,<:Any})

Creates a fleet, a logical group of robots running the same robot application.

Arguments

name: The name of the fleet.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"tags": A map that contains tag keys and tag values that are attached to the fleet.

Main.Robomaker.create_robot — Method

create_robot(architecture, greengrass_group_id, name)
create_robot(architecture, greengrass_group_id, name, params::Dict{String,<:Any})

Creates a robot.

Arguments

architecture: The target architecture of the robot.

greengrass_group_id: The Greengrass group id.

name: The name for the robot.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"tags": A map that contains tag keys and tag values that are attached to the robot.

Main.Robomaker.create_robot_application — Method

create_robot_application(name, robot_software_suite, sources)
create_robot_application(name, robot_software_suite, sources, params::Dict{String,<:Any})

Creates a robot application.

Arguments

name: The name of the robot application.

robot_software_suite: The robot software suite (ROS distribuition) used by the robot application.

sources: The sources of the robot application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"tags": A map that contains tag keys and tag values that are attached to the robot application.

Main.Robomaker.create_robot_application_version — Method

create_robot_application_version(application)
create_robot_application_version(application, params::Dict{String,<:Any})

Creates a version of a robot application.

Arguments

application: The application information for the robot application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"currentRevisionId": The current revision id for the robot application. If you provide a value and it matches the latest revision ID, a new version will be created.

Main.Robomaker.create_simulation_application — Method

create_simulation_application(name, robot_software_suite, simulation_software_suite, sources)
create_simulation_application(name, robot_software_suite, simulation_software_suite, sources, params::Dict{String,<:Any})

Creates a simulation application.

Arguments

name: The name of the simulation application.

robot_software_suite: The robot software suite (ROS distribution) used by the simulation application.

simulation_software_suite: The simulation software suite used by the simulation application.

sources: The sources of the simulation application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"renderingEngine": The rendering engine for the simulation application.

"tags": A map that contains tag keys and tag values that are attached to the simulation application.

Main.Robomaker.create_simulation_application_version — Method

create_simulation_application_version(application)
create_simulation_application_version(application, params::Dict{String,<:Any})

Creates a simulation application with a specific revision id.

Arguments

application: The application information for the simulation application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"currentRevisionId": The current revision id for the simulation application. If you provide a value and it matches the latest revision ID, a new version will be created.

Main.Robomaker.create_simulation_job — Method

create_simulation_job(iam_role, max_job_duration_in_seconds)
create_simulation_job(iam_role, max_job_duration_in_seconds, params::Dict{String,<:Any})

Creates a simulation job. After 90 days, simulation jobs expire and will be deleted. They will no longer be accessible.

Arguments

iam_role: The IAM role name that allows the simulation instance to call the AWS APIs that are specified in its associated policies on your behalf. This is how credentials are passed in to your simulation job.

max_job_duration_in_seconds: The maximum simulation job duration in seconds (up to 14 days or 1,209,600 seconds. When maxJobDurationInSeconds is reached, the simulation job will status will transition to Completed.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientRequestToken": Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

"compute": Compute information for the simulation job.

"dataSources": Specify data sources to mount read-only files from S3 into your simulation. These files are available under /opt/robomaker/datasources/datasourcename. There is a limit of 100 files and a combined size of 25GB for all DataSourceConfig objects.

"failureBehavior": The failure behavior the simulation job. Continue Restart the simulation job in the same host instance. Fail Stop the simulation job and terminate the instance.

"loggingConfig": The logging configuration.

"outputLocation": Location for output files generated by the simulation job.

"robotApplications": The robot application to use in the simulation job.

"simulationApplications": The simulation application to use in the simulation job.

"tags": A map that contains tag keys and tag values that are attached to the simulation job.

"vpcConfig": If your simulation job accesses resources in a VPC, you provide this parameter identifying the list of security group IDs and subnet IDs. These must belong to the same VPC. You must provide at least one security group and one subnet ID.

Main.Robomaker.create_world_export_job — Method

create_world_export_job(iam_role, output_location, worlds)
create_world_export_job(iam_role, output_location, worlds, params::Dict{String,<:Any})

Creates a world export job.

Arguments

iam_role: The IAM role that the world export process uses to access the Amazon S3 bucket and put the export.

output_location:

worlds: A list of Amazon Resource Names (arns) that correspond to worlds to export.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientRequestToken": Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

"tags": A map that contains tag keys and tag values that are attached to the world export job.

Main.Robomaker.create_world_generation_job — Method

create_world_generation_job(template, world_count)
create_world_generation_job(template, world_count, params::Dict{String,<:Any})

Creates worlds using the specified template.

Arguments

template: The Amazon Resource Name (arn) of the world template describing the worlds you want to create.

world_count: Information about the world count.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientRequestToken": Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

"tags": A map that contains tag keys and tag values that are attached to the world generator job.

"worldTags": A map that contains tag keys and tag values that are attached to the generated worlds.

Main.Robomaker.create_world_template — Method

create_world_template()
create_world_template(params::Dict{String,<:Any})

Creates a world template.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientRequestToken": Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

"name": The name of the world template.

"tags": A map that contains tag keys and tag values that are attached to the world template.

"templateBody": The world template body.

"templateLocation": The location of the world template.

Main.Robomaker.delete_fleet — Method

delete_fleet(fleet)
delete_fleet(fleet, params::Dict{String,<:Any})

Deletes a fleet.

Arguments

fleet: The Amazon Resource Name (ARN) of the fleet.

Main.Robomaker.delete_robot — Method

delete_robot(robot)
delete_robot(robot, params::Dict{String,<:Any})

Deletes a robot.

Arguments

robot: The Amazon Resource Name (ARN) of the robot.

Main.Robomaker.delete_robot_application — Method

delete_robot_application(application)
delete_robot_application(application, params::Dict{String,<:Any})

Deletes a robot application.

Arguments

application: The Amazon Resource Name (ARN) of the the robot application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"applicationVersion": The version of the robot application to delete.

Main.Robomaker.delete_simulation_application — Method

delete_simulation_application(application)
delete_simulation_application(application, params::Dict{String,<:Any})

Deletes a simulation application.

Arguments

application: The application information for the simulation application to delete.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"applicationVersion": The version of the simulation application to delete.

Main.Robomaker.delete_world_template — Method

delete_world_template(template)
delete_world_template(template, params::Dict{String,<:Any})

Deletes a world template.

Arguments

template: The Amazon Resource Name (arn) of the world template you want to delete.

Main.Robomaker.deregister_robot — Method

deregister_robot(fleet, robot)
deregister_robot(fleet, robot, params::Dict{String,<:Any})

Deregisters a robot.

Arguments

fleet: The Amazon Resource Name (ARN) of the fleet.

robot: The Amazon Resource Name (ARN) of the robot.

Main.Robomaker.describe_deployment_job — Method

describe_deployment_job(job)
describe_deployment_job(job, params::Dict{String,<:Any})

Describes a deployment job.

Arguments

job: The Amazon Resource Name (ARN) of the deployment job.

Main.Robomaker.describe_fleet — Method

describe_fleet(fleet)
describe_fleet(fleet, params::Dict{String,<:Any})

Describes a fleet.

Arguments

fleet: The Amazon Resource Name (ARN) of the fleet.

Main.Robomaker.describe_robot — Method

describe_robot(robot)
describe_robot(robot, params::Dict{String,<:Any})

Describes a robot.

Arguments

robot: The Amazon Resource Name (ARN) of the robot to be described.

Main.Robomaker.describe_robot_application — Method

describe_robot_application(application)
describe_robot_application(application, params::Dict{String,<:Any})

Describes a robot application.

Arguments

application: The Amazon Resource Name (ARN) of the robot application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"applicationVersion": The version of the robot application to describe.

Main.Robomaker.describe_simulation_application — Method

describe_simulation_application(application)
describe_simulation_application(application, params::Dict{String,<:Any})

Describes a simulation application.

Arguments

application: The application information for the simulation application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"applicationVersion": The version of the simulation application to describe.

Main.Robomaker.describe_simulation_job — Method

describe_simulation_job(job)
describe_simulation_job(job, params::Dict{String,<:Any})

Describes a simulation job.

Arguments

job: The Amazon Resource Name (ARN) of the simulation job to be described.

Main.Robomaker.describe_simulation_job_batch — Method

describe_simulation_job_batch(batch)
describe_simulation_job_batch(batch, params::Dict{String,<:Any})

Describes a simulation job batch.

Arguments

batch: The id of the batch to describe.

Main.Robomaker.describe_world — Method

describe_world(world)
describe_world(world, params::Dict{String,<:Any})

Describes a world.

Arguments

world: The Amazon Resource Name (arn) of the world you want to describe.

Main.Robomaker.describe_world_export_job — Method

describe_world_export_job(job)
describe_world_export_job(job, params::Dict{String,<:Any})

Describes a world export job.

Arguments

job: The Amazon Resource Name (arn) of the world export job to describe.

Main.Robomaker.describe_world_generation_job — Method

describe_world_generation_job(job)
describe_world_generation_job(job, params::Dict{String,<:Any})

Describes a world generation job.

Arguments

job: The Amazon Resource Name (arn) of the world generation job to describe.

Main.Robomaker.describe_world_template — Method

describe_world_template(template)
describe_world_template(template, params::Dict{String,<:Any})

Describes a world template.

Arguments

template: The Amazon Resource Name (arn) of the world template you want to describe.

Main.Robomaker.get_world_template_body — Method

get_world_template_body()
get_world_template_body(params::Dict{String,<:Any})

Gets the world template body.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"generationJob": The Amazon Resource Name (arn) of the world generator job.

"template": The Amazon Resource Name (arn) of the world template.

Main.Robomaker.list_deployment_jobs — Method

list_deployment_jobs()
list_deployment_jobs(params::Dict{String,<:Any})

Returns a list of deployment jobs for a fleet. You can optionally provide filters to retrieve specific deployment jobs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filters": Optional filters to limit results. The filter names status and fleetName are supported. When filtering, you must use the complete value of the filtered item. You can use up to three filters, but they must be for the same named item. For example, if you are looking for items with the status InProgress or the status Pending.

"maxResults": When this parameter is used, ListDeploymentJobs only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListDeploymentJobs request with the returned nextToken value. This value can be between 1 and 200. If this parameter is not used, then ListDeploymentJobs returns up to 200 results and a nextToken value if applicable.

"nextToken": If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListDeploymentJobs again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

Main.Robomaker.list_fleets — Method

list_fleets()
list_fleets(params::Dict{String,<:Any})

Returns a list of fleets. You can optionally provide filters to retrieve specific fleets.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filters": Optional filters to limit results. The filter name name is supported. When filtering, you must use the complete value of the filtered item. You can use up to three filters.

"maxResults": When this parameter is used, ListFleets only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListFleets request with the returned nextToken value. This value can be between 1 and 200. If this parameter is not used, then ListFleets returns up to 200 results and a nextToken value if applicable.

"nextToken": If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListFleets again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null. This token should be treated as an opaque identifier that is only used to retrieve the next items in a list and not for other programmatic purposes.

Main.Robomaker.list_robot_applications — Method

list_robot_applications()
list_robot_applications(params::Dict{String,<:Any})

Returns a list of robot application. You can optionally provide filters to retrieve specific robot applications.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filters": Optional filters to limit results. The filter name name is supported. When filtering, you must use the complete value of the filtered item. You can use up to three filters.

"maxResults": When this parameter is used, ListRobotApplications only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListRobotApplications request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListRobotApplications returns up to 100 results and a nextToken value if applicable.

"nextToken": If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListRobotApplications again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

"versionQualifier": The version qualifier of the robot application.

Main.Robomaker.list_robots — Method

list_robots()
list_robots(params::Dict{String,<:Any})

Returns a list of robots. You can optionally provide filters to retrieve specific robots.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filters": Optional filters to limit results. The filter names status and fleetName are supported. When filtering, you must use the complete value of the filtered item. You can use up to three filters, but they must be for the same named item. For example, if you are looking for items with the status Registered or the status Available.

"maxResults": When this parameter is used, ListRobots only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListRobots request with the returned nextToken value. This value can be between 1 and 200. If this parameter is not used, then ListRobots returns up to 200 results and a nextToken value if applicable.

"nextToken": If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListRobots again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

Main.Robomaker.list_simulation_applications — Method

list_simulation_applications()
list_simulation_applications(params::Dict{String,<:Any})

Returns a list of simulation applications. You can optionally provide filters to retrieve specific simulation applications.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filters": Optional list of filters to limit results. The filter name name is supported. When filtering, you must use the complete value of the filtered item. You can use up to three filters.

"maxResults": When this parameter is used, ListSimulationApplications only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListSimulationApplications request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListSimulationApplications returns up to 100 results and a nextToken value if applicable.

"nextToken": If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListSimulationApplications again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

"versionQualifier": The version qualifier of the simulation application.

Main.Robomaker.list_simulation_job_batches — Method

list_simulation_job_batches()
list_simulation_job_batches(params::Dict{String,<:Any})

Returns a list simulation job batches. You can optionally provide filters to retrieve specific simulation batch jobs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filters": Optional filters to limit results.

"maxResults": When this parameter is used, ListSimulationJobBatches only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListSimulationJobBatches request with the returned nextToken value.

"nextToken": If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListSimulationJobBatches again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

Main.Robomaker.list_simulation_jobs — Method

list_simulation_jobs()
list_simulation_jobs(params::Dict{String,<:Any})

Returns a list of simulation jobs. You can optionally provide filters to retrieve specific simulation jobs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filters": Optional filters to limit results. The filter names status and simulationApplicationName and robotApplicationName are supported. When filtering, you must use the complete value of the filtered item. You can use up to three filters, but they must be for the same named item. For example, if you are looking for items with the status Preparing or the status Running.

"maxResults": When this parameter is used, ListSimulationJobs only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListSimulationJobs request with the returned nextToken value. This value can be between 1 and 1000. If this parameter is not used, then ListSimulationJobs returns up to 1000 results and a nextToken value if applicable.

"nextToken": If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListSimulationJobs again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

Main.Robomaker.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists all tags on a AWS RoboMaker resource.

Arguments

resource_arn: The AWS RoboMaker Amazon Resource Name (ARN) with tags to be listed.

Main.Robomaker.list_world_export_jobs — Method

list_world_export_jobs()
list_world_export_jobs(params::Dict{String,<:Any})

Lists world export jobs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filters": Optional filters to limit results. You can use generationJobId and templateId.

"maxResults": When this parameter is used, ListWorldExportJobs only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListWorldExportJobs request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListWorldExportJobs returns up to 100 results and a nextToken value if applicable.

"nextToken": If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListWorldExportJobs again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

Main.Robomaker.list_world_generation_jobs — Method

list_world_generation_jobs()
list_world_generation_jobs(params::Dict{String,<:Any})

Lists world generator jobs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filters": Optional filters to limit results. You can use status and templateId.

"maxResults": When this parameter is used, ListWorldGeneratorJobs only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListWorldGeneratorJobs request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListWorldGeneratorJobs returns up to 100 results and a nextToken value if applicable.

"nextToken": If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListWorldGenerationJobsRequest again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

Main.Robomaker.list_world_templates — Method

list_world_templates()
list_world_templates(params::Dict{String,<:Any})

Lists world templates.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": When this parameter is used, ListWorldTemplates only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListWorldTemplates request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListWorldTemplates returns up to 100 results and a nextToken value if applicable.

"nextToken": If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListWorldTemplates again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

Main.Robomaker.list_worlds — Method

list_worlds()
list_worlds(params::Dict{String,<:Any})

Lists worlds.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filters": Optional filters to limit results. You can use status.

"maxResults": When this parameter is used, ListWorlds only returns maxResults results in a single page along with a nextToken response element. The remaining results of the initial request can be seen by sending another ListWorlds request with the returned nextToken value. This value can be between 1 and 100. If this parameter is not used, then ListWorlds returns up to 100 results and a nextToken value if applicable.

"nextToken": If the previous paginated request did not return all of the remaining results, the response object's nextToken parameter value is set to a token. To retrieve the next set of results, call ListWorlds again and assign that token to the request object's nextToken parameter. If there are no remaining results, the previous response object's NextToken parameter is set to null.

Main.Robomaker.register_robot — Method

register_robot(fleet, robot)
register_robot(fleet, robot, params::Dict{String,<:Any})

Registers a robot with a fleet.

Arguments

fleet: The Amazon Resource Name (ARN) of the fleet.

robot: The Amazon Resource Name (ARN) of the robot.

Main.Robomaker.restart_simulation_job — Method

restart_simulation_job(job)
restart_simulation_job(job, params::Dict{String,<:Any})

Restarts a running simulation job.

Arguments

job: The Amazon Resource Name (ARN) of the simulation job.

Main.Robomaker.start_simulation_job_batch — Method

start_simulation_job_batch(create_simulation_job_requests)
start_simulation_job_batch(create_simulation_job_requests, params::Dict{String,<:Any})

Starts a new simulation job batch. The batch is defined using one or more SimulationJobRequest objects.

Arguments

create_simulation_job_requests: A list of simulation job requests to create in the batch.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"batchPolicy": The batch policy.

"clientRequestToken": Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

"tags": A map that contains tag keys and tag values that are attached to the deployment job batch.

Main.Robomaker.sync_deployment_job — Method

sync_deployment_job(client_request_token, fleet)
sync_deployment_job(client_request_token, fleet, params::Dict{String,<:Any})

Syncrhonizes robots in a fleet to the latest deployment. This is helpful if robots were added after a deployment.

Arguments

client_request_token: Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

fleet: The target fleet for the synchronization.

Main.Robomaker.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Adds or edits tags for a AWS RoboMaker resource. Each tag consists of a tag key and a tag value. Tag keys and tag values are both required, but tag values can be empty strings. For information about the rules that apply to tag keys and tag values, see User-Defined Tag Restrictions in the AWS Billing and Cost Management User Guide.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the AWS RoboMaker resource you are tagging.

tags: A map that contains tag keys and tag values that are attached to the resource.

Main.Robomaker.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes the specified tags from the specified AWS RoboMaker resource. To remove a tag, specify the tag key. To change the tag value of an existing tag key, use TagResource .

Arguments

resource_arn: The Amazon Resource Name (ARN) of the AWS RoboMaker resource you are removing tags.

tag_keys: A map that contains tag keys and tag values that will be unattached from the resource.

Main.Robomaker.update_robot_application — Method

update_robot_application(application, robot_software_suite, sources)
update_robot_application(application, robot_software_suite, sources, params::Dict{String,<:Any})

Updates a robot application.

Arguments

application: The application information for the robot application.

robot_software_suite: The robot software suite (ROS distribution) used by the robot application.

sources: The sources of the robot application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"currentRevisionId": The revision id for the robot application.

Main.Robomaker.update_simulation_application — Method

update_simulation_application(application, robot_software_suite, simulation_software_suite, sources)
update_simulation_application(application, robot_software_suite, simulation_software_suite, sources, params::Dict{String,<:Any})

Updates a simulation application.

Arguments

application: The application information for the simulation application.

robot_software_suite: Information about the robot software suite (ROS distribution).

simulation_software_suite: The simulation software suite used by the simulation application.

sources: The sources of the simulation application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"currentRevisionId": The revision id for the robot application.

"renderingEngine": The rendering engine for the simulation application.

Main.Robomaker.update_world_template — Method

update_world_template(template)
update_world_template(template, params::Dict{String,<:Any})

Updates a world template.

Arguments

template: The Amazon Resource Name (arn) of the world template to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"name": The name of the template.

"templateBody": The world template body.

"templateLocation": The location of the world template.