Route53 Recovery Control Config

This page documents function available when using the Route53_Recovery_Control_Config module, created with @service Route53_Recovery_Control_Config.

Index

Main.Route53_Recovery_Control_Config.create_cluster

Main.Route53_Recovery_Control_Config.create_control_panel

Main.Route53_Recovery_Control_Config.create_routing_control

Main.Route53_Recovery_Control_Config.create_safety_rule

Main.Route53_Recovery_Control_Config.delete_cluster

Main.Route53_Recovery_Control_Config.delete_control_panel

Main.Route53_Recovery_Control_Config.delete_routing_control

Main.Route53_Recovery_Control_Config.delete_safety_rule

Main.Route53_Recovery_Control_Config.describe_cluster

Main.Route53_Recovery_Control_Config.describe_control_panel

Main.Route53_Recovery_Control_Config.describe_routing_control

Main.Route53_Recovery_Control_Config.describe_safety_rule

Main.Route53_Recovery_Control_Config.get_resource_policy

Main.Route53_Recovery_Control_Config.list_associated_route53_health_checks

Main.Route53_Recovery_Control_Config.list_clusters

Main.Route53_Recovery_Control_Config.list_control_panels

Main.Route53_Recovery_Control_Config.list_routing_controls

Main.Route53_Recovery_Control_Config.list_safety_rules

Main.Route53_Recovery_Control_Config.list_tags_for_resource

Main.Route53_Recovery_Control_Config.tag_resource

Main.Route53_Recovery_Control_Config.untag_resource

Main.Route53_Recovery_Control_Config.update_control_panel

Main.Route53_Recovery_Control_Config.update_routing_control

Main.Route53_Recovery_Control_Config.update_safety_rule

Documentation

Main.Route53_Recovery_Control_Config.create_cluster — Method

create_cluster(cluster_name)
create_cluster(cluster_name, params::Dict{String,<:Any})

Create a new cluster. A cluster is a set of redundant Regional endpoints against which you can run API calls to update or get the state of one or more routing controls. Each cluster has a name, status, Amazon Resource Name (ARN), and an array of the five cluster endpoints (one for each supported Amazon Web Services Region) that you can use with API calls to the cluster data plane.

Arguments

cluster_name: The name of the cluster.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": A unique, case-sensitive string of up to 64 ASCII characters. To make an idempotent API request with an action, specify a client token in the request.

"Tags": The tags associated with the cluster.

Main.Route53_Recovery_Control_Config.create_control_panel — Method

create_control_panel(cluster_arn, control_panel_name)
create_control_panel(cluster_arn, control_panel_name, params::Dict{String,<:Any})

Creates a new control panel. A control panel represents a group of routing controls that can be changed together in a single transaction. You can use a control panel to centrally view the operational status of applications across your organization, and trigger multi-app failovers in a single transaction, for example, to fail over an Availability Zone or Amazon Web Services Region.

Arguments

cluster_arn: The Amazon Resource Name (ARN) of the cluster for the control panel.

control_panel_name: The name of the control panel.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": A unique, case-sensitive string of up to 64 ASCII characters. To make an idempotent API request with an action, specify a client token in the request.

"Tags": The tags associated with the control panel.

Main.Route53_Recovery_Control_Config.create_routing_control — Method

create_routing_control(cluster_arn, routing_control_name)
create_routing_control(cluster_arn, routing_control_name, params::Dict{String,<:Any})

Creates a new routing control. A routing control has one of two states: ON and OFF. You can map the routing control state to the state of an Amazon Route 53 health check, which can be used to control traffic routing. To get or update the routing control state, see the Recovery Cluster (data plane) API actions for Amazon Route 53 Application Recovery Controller.

Arguments

cluster_arn: The Amazon Resource Name (ARN) of the cluster that includes the routing control.

routing_control_name: The name of the routing control.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": A unique, case-sensitive string of up to 64 ASCII characters. To make an idempotent API request with an action, specify a client token in the request.

"ControlPanelArn": The Amazon Resource Name (ARN) of the control panel that includes the routing control.

Main.Route53_Recovery_Control_Config.create_safety_rule — Method

create_safety_rule()
create_safety_rule(params::Dict{String,<:Any})

Creates a safety rule in a control panel. Safety rules let you add safeguards around changing routing control states, and for enabling and disabling routing controls, to help prevent unexpected outcomes. There are two types of safety rules: assertion rules and gating rules. Assertion rule: An assertion rule enforces that, when you change a routing control state, that a certain criteria is met. For example, the criteria might be that at least one routing control state is On after the transaction so that traffic continues to flow to at least one cell for the application. This ensures that you avoid a fail-open scenario. Gating rule: A gating rule lets you configure a gating routing control as an overall "on/off" switch for a group of routing controls. Or, you can configure more complex gating scenarios, for example by configuring multiple gating routing controls. For more information, see Safety rules in the Amazon Route 53 Application Recovery Controller Developer Guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AssertionRule": The assertion rule requested.

"ClientToken": A unique, case-sensitive string of up to 64 ASCII characters. To make an idempotent API request with an action, specify a client token in the request.

"GatingRule": The gating rule requested.

"Tags": The tags associated with the safety rule.

Main.Route53_Recovery_Control_Config.delete_cluster — Method

delete_cluster(cluster_arn)
delete_cluster(cluster_arn, params::Dict{String,<:Any})

Delete a cluster.

Arguments

cluster_arn: The Amazon Resource Name (ARN) of the cluster that you're deleting.

Main.Route53_Recovery_Control_Config.delete_control_panel — Method

delete_control_panel(control_panel_arn)
delete_control_panel(control_panel_arn, params::Dict{String,<:Any})

Deletes a control panel.

Arguments

control_panel_arn: The Amazon Resource Name (ARN) of the control panel.

Main.Route53_Recovery_Control_Config.delete_routing_control — Method

delete_routing_control(routing_control_arn)
delete_routing_control(routing_control_arn, params::Dict{String,<:Any})

Deletes a routing control.

Arguments

routing_control_arn: The Amazon Resource Name (ARN) of the routing control that you're deleting.

Main.Route53_Recovery_Control_Config.delete_safety_rule — Method

delete_safety_rule(safety_rule_arn)
delete_safety_rule(safety_rule_arn, params::Dict{String,<:Any})

Deletes a safety rule./&gt;

Arguments

safety_rule_arn: The ARN of the safety rule.

Main.Route53_Recovery_Control_Config.describe_cluster — Method

describe_cluster(cluster_arn)
describe_cluster(cluster_arn, params::Dict{String,<:Any})

Display the details about a cluster. The response includes the cluster name, endpoints, status, and Amazon Resource Name (ARN).

Arguments

cluster_arn: The Amazon Resource Name (ARN) of the cluster.

Main.Route53_Recovery_Control_Config.describe_control_panel — Method

describe_control_panel(control_panel_arn)
describe_control_panel(control_panel_arn, params::Dict{String,<:Any})

Displays details about a control panel.

Arguments

control_panel_arn: The Amazon Resource Name (ARN) of the control panel.

Main.Route53_Recovery_Control_Config.describe_routing_control — Method

describe_routing_control(routing_control_arn)
describe_routing_control(routing_control_arn, params::Dict{String,<:Any})

Displays details about a routing control. A routing control has one of two states: ON and OFF. You can map the routing control state to the state of an Amazon Route 53 health check, which can be used to control routing. To get or update the routing control state, see the Recovery Cluster (data plane) API actions for Amazon Route 53 Application Recovery Controller.

Arguments

routing_control_arn: The Amazon Resource Name (ARN) of the routing control.

Main.Route53_Recovery_Control_Config.describe_safety_rule — Method

describe_safety_rule(safety_rule_arn)
describe_safety_rule(safety_rule_arn, params::Dict{String,<:Any})

Returns information about a safety rule.

Arguments

safety_rule_arn: The ARN of the safety rule.

Main.Route53_Recovery_Control_Config.get_resource_policy — Method

get_resource_policy(resource_arn)
get_resource_policy(resource_arn, params::Dict{String,<:Any})

Get information about the resource policy for a cluster.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

Main.Route53_Recovery_Control_Config.list_associated_route53_health_checks — Method

list_associated_route53_health_checks(routing_control_arn)
list_associated_route53_health_checks(routing_control_arn, params::Dict{String,<:Any})

Returns an array of all Amazon Route 53 health checks associated with a specific routing control.

Arguments

routing_control_arn: The Amazon Resource Name (ARN) of the routing control.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The number of objects that you want to return with this call.

"NextToken": The token that identifies which batch of results you want to see.

Main.Route53_Recovery_Control_Config.list_clusters — Method

list_clusters()
list_clusters(params::Dict{String,<:Any})

Returns an array of all the clusters in an account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The number of objects that you want to return with this call.

"NextToken": The token that identifies which batch of results you want to see.

Main.Route53_Recovery_Control_Config.list_control_panels — Method

list_control_panels()
list_control_panels(params::Dict{String,<:Any})

Returns an array of control panels in an account or in a cluster.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClusterArn": The Amazon Resource Name (ARN) of a cluster.

"MaxResults": The number of objects that you want to return with this call.

"NextToken": The token that identifies which batch of results you want to see.

Main.Route53_Recovery_Control_Config.list_routing_controls — Method

list_routing_controls(control_panel_arn)
list_routing_controls(control_panel_arn, params::Dict{String,<:Any})

Returns an array of routing controls for a control panel. A routing control is an Amazon Route 53 Application Recovery Controller construct that has one of two states: ON and OFF. You can map the routing control state to the state of an Amazon Route 53 health check, which can be used to control routing.

Arguments

control_panel_arn: The Amazon Resource Name (ARN) of the control panel.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The number of objects that you want to return with this call.

"NextToken": The token that identifies which batch of results you want to see.

Main.Route53_Recovery_Control_Config.list_safety_rules — Method

list_safety_rules(control_panel_arn)
list_safety_rules(control_panel_arn, params::Dict{String,<:Any})

List the safety rules (the assertion rules and gating rules) that you've defined for the routing controls in a control panel.

Arguments

control_panel_arn: The Amazon Resource Name (ARN) of the control panel.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The number of objects that you want to return with this call.

"NextToken": The token that identifies which batch of results you want to see.

Main.Route53_Recovery_Control_Config.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists the tags for a resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) for the resource that's tagged.

Main.Route53_Recovery_Control_Config.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Adds a tag to a resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) for the resource that's tagged.

tags: The tags associated with the resource.

Main.Route53_Recovery_Control_Config.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes a tag from a resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) for the resource that's tagged.

tag_keys: Keys for the tags to be removed.

Main.Route53_Recovery_Control_Config.update_control_panel — Method

update_control_panel(control_panel_arn, control_panel_name)
update_control_panel(control_panel_arn, control_panel_name, params::Dict{String,<:Any})

Updates a control panel. The only update you can make to a control panel is to change the name of the control panel.

Arguments

control_panel_arn: The Amazon Resource Name (ARN) of the control panel.

control_panel_name: The name of the control panel.

Main.Route53_Recovery_Control_Config.update_routing_control — Method

update_routing_control(routing_control_arn, routing_control_name)
update_routing_control(routing_control_arn, routing_control_name, params::Dict{String,<:Any})

Updates a routing control. You can only update the name of the routing control. To get or update the routing control state, see the Recovery Cluster (data plane) API actions for Amazon Route 53 Application Recovery Controller.

Arguments

routing_control_arn: The Amazon Resource Name (ARN) of the routing control.

routing_control_name: The name of the routing control.

Main.Route53_Recovery_Control_Config.update_safety_rule — Method

update_safety_rule()
update_safety_rule(params::Dict{String,<:Any})

Update a safety rule (an assertion rule or gating rule). You can only update the name and the waiting period for a safety rule. To make other updates, delete the safety rule and create a new one.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AssertionRuleUpdate": The assertion rule to update.

"GatingRuleUpdate": The gating rule to update.