Route53 Recovery Readiness

This page documents function available when using the Route53_Recovery_Readiness module, created with @service Route53_Recovery_Readiness.

Index

Main.Route53_Recovery_Readiness.create_cell

Main.Route53_Recovery_Readiness.create_cross_account_authorization

Main.Route53_Recovery_Readiness.create_readiness_check

Main.Route53_Recovery_Readiness.create_recovery_group

Main.Route53_Recovery_Readiness.create_resource_set

Main.Route53_Recovery_Readiness.delete_cell

Main.Route53_Recovery_Readiness.delete_cross_account_authorization

Main.Route53_Recovery_Readiness.delete_readiness_check

Main.Route53_Recovery_Readiness.delete_recovery_group

Main.Route53_Recovery_Readiness.delete_resource_set

Main.Route53_Recovery_Readiness.get_architecture_recommendations

Main.Route53_Recovery_Readiness.get_cell

Main.Route53_Recovery_Readiness.get_cell_readiness_summary

Main.Route53_Recovery_Readiness.get_readiness_check

Main.Route53_Recovery_Readiness.get_readiness_check_resource_status

Main.Route53_Recovery_Readiness.get_readiness_check_status

Main.Route53_Recovery_Readiness.get_recovery_group

Main.Route53_Recovery_Readiness.get_recovery_group_readiness_summary

Main.Route53_Recovery_Readiness.get_resource_set

Main.Route53_Recovery_Readiness.list_cells

Main.Route53_Recovery_Readiness.list_cross_account_authorizations

Main.Route53_Recovery_Readiness.list_readiness_checks

Main.Route53_Recovery_Readiness.list_recovery_groups

Main.Route53_Recovery_Readiness.list_resource_sets

Main.Route53_Recovery_Readiness.list_rules

Main.Route53_Recovery_Readiness.list_tags_for_resources

Main.Route53_Recovery_Readiness.tag_resource

Main.Route53_Recovery_Readiness.untag_resource

Main.Route53_Recovery_Readiness.update_cell

Main.Route53_Recovery_Readiness.update_readiness_check

Main.Route53_Recovery_Readiness.update_recovery_group

Main.Route53_Recovery_Readiness.update_resource_set

Documentation

Main.Route53_Recovery_Readiness.create_cell — Method

create_cell(cell_name)
create_cell(cell_name, params::Dict{String,<:Any})

Creates a cell in an account.

Arguments

cell_name: The name of the cell to create.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"cells": A list of cell Amazon Resource Names (ARNs) contained within this cell, for use in nested cells. For example, Availability Zones within specific Amazon Web Services Regions.

"tags":

Main.Route53_Recovery_Readiness.create_cross_account_authorization — Method

create_cross_account_authorization(cross_account_authorization)
create_cross_account_authorization(cross_account_authorization, params::Dict{String,<:Any})

Creates a cross-account readiness authorization. This lets you authorize another account to work with Route 53 Application Recovery Controller, for example, to check the readiness status of resources in a separate account.

Arguments

cross_account_authorization: The cross-account authorization.

Main.Route53_Recovery_Readiness.create_readiness_check — Method

create_readiness_check(readiness_check_name, resource_set_name)
create_readiness_check(readiness_check_name, resource_set_name, params::Dict{String,<:Any})

Creates a readiness check in an account. A readiness check monitors a resource set in your application, such as a set of Amazon Aurora instances, that Application Recovery Controller is auditing recovery readiness for. The audits run once every minute on every resource that's associated with a readiness check.

Arguments

readiness_check_name: The name of the readiness check to create.

resource_set_name: The name of the resource set to check.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"tags":

Main.Route53_Recovery_Readiness.create_recovery_group — Method

create_recovery_group(recovery_group_name)
create_recovery_group(recovery_group_name, params::Dict{String,<:Any})

Creates a recovery group in an account. A recovery group corresponds to an application and includes a list of the cells that make up the application.

Arguments

recovery_group_name: The name of the recovery group to create.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"cells": A list of the cell Amazon Resource Names (ARNs) in the recovery group.

"tags":

Main.Route53_Recovery_Readiness.create_resource_set — Method

create_resource_set(resource_set_name, resource_set_type, resources)
create_resource_set(resource_set_name, resource_set_type, resources, params::Dict{String,<:Any})

Creates a resource set. A resource set is a set of resources of one type that span multiple cells. You can associate a resource set with a readiness check to monitor the resources for failover readiness.

Arguments

resource_set_name: The name of the resource set to create.

resource_set_type: The resource type of the resources in the resource set. Enter one of the following values for resource type: AWS::ApiGateway::Stage, AWS::ApiGatewayV2::Stage, AWS::AutoScaling::AutoScalingGroup, AWS::CloudWatch::Alarm, AWS::EC2::CustomerGateway, AWS::DynamoDB::Table, AWS::EC2::Volume, AWS::ElasticLoadBalancing::LoadBalancer, AWS::ElasticLoadBalancingV2::LoadBalancer, AWS::Lambda::Function, AWS::MSK::Cluster, AWS::RDS::DBCluster, AWS::Route53::HealthCheck, AWS::SQS::Queue, AWS::SNS::Topic, AWS::SNS::Subscription, AWS::EC2::VPC, AWS::EC2::VPNConnection, AWS::EC2::VPNGateway, AWS::Route53RecoveryReadiness::DNSTargetResource

resources: A list of resource objects in the resource set.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"tags": A tag to associate with the parameters for a resource set.

Main.Route53_Recovery_Readiness.delete_cell — Method

delete_cell(cell_name)
delete_cell(cell_name, params::Dict{String,<:Any})

Delete a cell. When successful, the response code is 204, with no response body.

Arguments

cell_name: The name of the cell.

Main.Route53_Recovery_Readiness.delete_cross_account_authorization — Method

delete_cross_account_authorization(cross_account_authorization)
delete_cross_account_authorization(cross_account_authorization, params::Dict{String,<:Any})

Deletes cross account readiness authorization.

Arguments

cross_account_authorization: The cross-account authorization.

Main.Route53_Recovery_Readiness.delete_readiness_check — Method

delete_readiness_check(readiness_check_name)
delete_readiness_check(readiness_check_name, params::Dict{String,<:Any})

Deletes a readiness check.

Arguments

readiness_check_name: Name of a readiness check.

Main.Route53_Recovery_Readiness.delete_recovery_group — Method

delete_recovery_group(recovery_group_name)
delete_recovery_group(recovery_group_name, params::Dict{String,<:Any})

Deletes a recovery group.

Arguments

recovery_group_name: The name of a recovery group.

Main.Route53_Recovery_Readiness.delete_resource_set — Method

delete_resource_set(resource_set_name)
delete_resource_set(resource_set_name, params::Dict{String,<:Any})

Deletes a resource set.

Arguments

resource_set_name: Name of a resource set.

Main.Route53_Recovery_Readiness.get_architecture_recommendations — Method

get_architecture_recommendations(recovery_group_name)
get_architecture_recommendations(recovery_group_name, params::Dict{String,<:Any})

Gets recommendations about architecture designs for improving resiliency for an application, based on a recovery group.

Arguments

recovery_group_name: The name of a recovery group.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The number of objects that you want to return with this call.

"nextToken": The token that identifies which batch of results you want to see.

Main.Route53_Recovery_Readiness.get_cell — Method

get_cell(cell_name)
get_cell(cell_name, params::Dict{String,<:Any})

Gets information about a cell including cell name, cell Amazon Resource Name (ARN), ARNs of nested cells for this cell, and a list of those cell ARNs with their associated recovery group ARNs.

Arguments

cell_name: The name of the cell.

Main.Route53_Recovery_Readiness.get_cell_readiness_summary — Method

get_cell_readiness_summary(cell_name)
get_cell_readiness_summary(cell_name, params::Dict{String,<:Any})

Gets readiness for a cell. Aggregates the readiness of all the resources that are associated with the cell into a single value.

Arguments

cell_name: The name of the cell.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The number of objects that you want to return with this call.

"nextToken": The token that identifies which batch of results you want to see.

Main.Route53_Recovery_Readiness.get_readiness_check — Method

get_readiness_check(readiness_check_name)
get_readiness_check(readiness_check_name, params::Dict{String,<:Any})

Gets details about a readiness check.

Arguments

readiness_check_name: Name of a readiness check.

Main.Route53_Recovery_Readiness.get_readiness_check_resource_status — Method

get_readiness_check_resource_status(readiness_check_name, resource_identifier)
get_readiness_check_resource_status(readiness_check_name, resource_identifier, params::Dict{String,<:Any})

Gets individual readiness status for a readiness check. To see the overall readiness status for a recovery group, that considers the readiness status for all the readiness checks in the recovery group, use GetRecoveryGroupReadinessSummary.

Arguments

readiness_check_name: Name of a readiness check.

resource_identifier: The resource identifier, which is the Amazon Resource Name (ARN) or the identifier generated for the resource by Application Recovery Controller (for example, for a DNS target resource).

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The number of objects that you want to return with this call.

"nextToken": The token that identifies which batch of results you want to see.

Main.Route53_Recovery_Readiness.get_readiness_check_status — Method

get_readiness_check_status(readiness_check_name)
get_readiness_check_status(readiness_check_name, params::Dict{String,<:Any})

Gets the readiness status for an individual readiness check. To see the overall readiness status for a recovery group, that considers the readiness status for all the readiness checks in a recovery group, use GetRecoveryGroupReadinessSummary.

Arguments

readiness_check_name: Name of a readiness check.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The number of objects that you want to return with this call.

"nextToken": The token that identifies which batch of results you want to see.

Main.Route53_Recovery_Readiness.get_recovery_group — Method

get_recovery_group(recovery_group_name)
get_recovery_group(recovery_group_name, params::Dict{String,<:Any})

Gets details about a recovery group, including a list of the cells that are included in it.

Arguments

recovery_group_name: The name of a recovery group.

Main.Route53_Recovery_Readiness.get_recovery_group_readiness_summary — Method

get_recovery_group_readiness_summary(recovery_group_name)
get_recovery_group_readiness_summary(recovery_group_name, params::Dict{String,<:Any})

Displays a summary of information about a recovery group's readiness status. Includes the readiness checks for resources in the recovery group and the readiness status of each one.

Arguments

recovery_group_name: The name of a recovery group.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The number of objects that you want to return with this call.

"nextToken": The token that identifies which batch of results you want to see.

Main.Route53_Recovery_Readiness.get_resource_set — Method

get_resource_set(resource_set_name)
get_resource_set(resource_set_name, params::Dict{String,<:Any})

Displays the details about a resource set, including a list of the resources in the set.

Arguments

resource_set_name: Name of a resource set.

Main.Route53_Recovery_Readiness.list_cells — Method

list_cells()
list_cells(params::Dict{String,<:Any})

Lists the cells for an account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The number of objects that you want to return with this call.

"nextToken": The token that identifies which batch of results you want to see.

Main.Route53_Recovery_Readiness.list_cross_account_authorizations — Method

list_cross_account_authorizations()
list_cross_account_authorizations(params::Dict{String,<:Any})

Lists the cross-account readiness authorizations that are in place for an account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The number of objects that you want to return with this call.

"nextToken": The token that identifies which batch of results you want to see.

Main.Route53_Recovery_Readiness.list_readiness_checks — Method

list_readiness_checks()
list_readiness_checks(params::Dict{String,<:Any})

Lists the readiness checks for an account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The number of objects that you want to return with this call.

"nextToken": The token that identifies which batch of results you want to see.

Main.Route53_Recovery_Readiness.list_recovery_groups — Method

list_recovery_groups()
list_recovery_groups(params::Dict{String,<:Any})

Lists the recovery groups in an account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The number of objects that you want to return with this call.

"nextToken": The token that identifies which batch of results you want to see.

Main.Route53_Recovery_Readiness.list_resource_sets — Method

list_resource_sets()
list_resource_sets(params::Dict{String,<:Any})

Lists the resource sets in an account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The number of objects that you want to return with this call.

"nextToken": The token that identifies which batch of results you want to see.

Main.Route53_Recovery_Readiness.list_rules — Method

list_rules()
list_rules(params::Dict{String,<:Any})

Lists all readiness rules, or lists the readiness rules for a specific resource type.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The number of objects that you want to return with this call.

"nextToken": The token that identifies which batch of results you want to see.

"resourceType": The resource type that a readiness rule applies to.

Main.Route53_Recovery_Readiness.list_tags_for_resources — Method

list_tags_for_resources(resource-arn)
list_tags_for_resources(resource-arn, params::Dict{String,<:Any})

Lists the tags for a resource.

Arguments

resource-arn: The Amazon Resource Name (ARN) for a resource.

Main.Route53_Recovery_Readiness.tag_resource — Method

tag_resource(resource-arn, tags)
tag_resource(resource-arn, tags, params::Dict{String,<:Any})

Adds a tag to a resource.

Arguments

resource-arn: The Amazon Resource Name (ARN) for a resource.

tags:

Main.Route53_Recovery_Readiness.untag_resource — Method

untag_resource(resource-arn, tag_keys)
untag_resource(resource-arn, tag_keys, params::Dict{String,<:Any})

Removes a tag from a resource.

Arguments

resource-arn: The Amazon Resource Name (ARN) for a resource.

tag_keys: The keys for tags you add to resources.

Main.Route53_Recovery_Readiness.update_cell — Method

update_cell(cell_name, cells)
update_cell(cell_name, cells, params::Dict{String,<:Any})

Updates a cell to replace the list of nested cells with a new list of nested cells.

Arguments

cell_name: The name of the cell.

cells: A list of cell Amazon Resource Names (ARNs), which completely replaces the previous list.

Main.Route53_Recovery_Readiness.update_readiness_check — Method

update_readiness_check(readiness_check_name, resource_set_name)
update_readiness_check(readiness_check_name, resource_set_name, params::Dict{String,<:Any})

Updates a readiness check.

Arguments

readiness_check_name: Name of a readiness check.

resource_set_name: The name of the resource set to be checked.

Main.Route53_Recovery_Readiness.update_recovery_group — Method

update_recovery_group(cells, recovery_group_name)
update_recovery_group(cells, recovery_group_name, params::Dict{String,<:Any})

Updates a recovery group.

Arguments

cells: A list of cell Amazon Resource Names (ARNs). This list completely replaces the previous list.

recovery_group_name: The name of a recovery group.

Main.Route53_Recovery_Readiness.update_resource_set — Method

update_resource_set(resource_set_name, resource_set_type, resources)
update_resource_set(resource_set_name, resource_set_type, resources, params::Dict{String,<:Any})

Updates a resource set.

Arguments

resource_set_name: Name of a resource set.

resource_set_type: The resource type of the resources in the resource set. Enter one of the following values for resource type: AWS::ApiGateway::Stage, AWS::ApiGatewayV2::Stage, AWS::AutoScaling::AutoScalingGroup, AWS::CloudWatch::Alarm, AWS::EC2::CustomerGateway, AWS::DynamoDB::Table, AWS::EC2::Volume, AWS::ElasticLoadBalancing::LoadBalancer, AWS::ElasticLoadBalancingV2::LoadBalancer, AWS::Lambda::Function, AWS::MSK::Cluster, AWS::RDS::DBCluster, AWS::Route53::HealthCheck, AWS::SQS::Queue, AWS::SNS::Topic, AWS::SNS::Subscription, AWS::EC2::VPC, AWS::EC2::VPNConnection, AWS::EC2::VPNGateway, AWS::Route53RecoveryReadiness::DNSTargetResource

resources: A list of resource objects.