S3Tables

This page documents function available when using the S3Tables module, created with @service S3Tables.

Index

Main.S3Tables.create_namespace

Main.S3Tables.create_table

Main.S3Tables.create_table_bucket

Main.S3Tables.delete_namespace

Main.S3Tables.delete_table

Main.S3Tables.delete_table_bucket

Main.S3Tables.delete_table_bucket_encryption

Main.S3Tables.delete_table_bucket_metrics_configuration

Main.S3Tables.delete_table_bucket_policy

Main.S3Tables.delete_table_bucket_replication

Main.S3Tables.delete_table_policy

Main.S3Tables.delete_table_replication

Main.S3Tables.get_namespace

Main.S3Tables.get_table

Main.S3Tables.get_table_bucket

Main.S3Tables.get_table_bucket_encryption

Main.S3Tables.get_table_bucket_maintenance_configuration

Main.S3Tables.get_table_bucket_metrics_configuration

Main.S3Tables.get_table_bucket_policy

Main.S3Tables.get_table_bucket_replication

Main.S3Tables.get_table_bucket_storage_class

Main.S3Tables.get_table_encryption

Main.S3Tables.get_table_maintenance_configuration

Main.S3Tables.get_table_maintenance_job_status

Main.S3Tables.get_table_metadata_location

Main.S3Tables.get_table_policy

Main.S3Tables.get_table_record_expiration_configuration

Main.S3Tables.get_table_record_expiration_job_status

Main.S3Tables.get_table_replication

Main.S3Tables.get_table_replication_status

Main.S3Tables.get_table_storage_class

Main.S3Tables.list_namespaces

Main.S3Tables.list_table_buckets

Main.S3Tables.list_tables

Main.S3Tables.list_tags_for_resource

Main.S3Tables.put_table_bucket_encryption

Main.S3Tables.put_table_bucket_maintenance_configuration

Main.S3Tables.put_table_bucket_metrics_configuration

Main.S3Tables.put_table_bucket_policy

Main.S3Tables.put_table_bucket_replication

Main.S3Tables.put_table_bucket_storage_class

Main.S3Tables.put_table_maintenance_configuration

Main.S3Tables.put_table_policy

Main.S3Tables.put_table_record_expiration_configuration

Main.S3Tables.put_table_replication

Main.S3Tables.rename_table

Main.S3Tables.tag_resource

Main.S3Tables.untag_resource

Main.S3Tables.update_table_metadata_location

Documentation

Main.S3Tables.create_namespace — Function

create_namespace(namespace, table_bucket_arn)
create_namespace(namespace, table_bucket_arn, params::Dict{String,<:Any})

Creates a namespace. A namespace is a logical grouping of tables within your table bucket, which you can use to organize tables. For more information, see Create a namespace in the Amazon Simple Storage Service User Guide.

Permissions

You must have the s3tables:CreateNamespace permission to use this operation.

Arguments

namespace: A name for the namespace.

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket to create the namespace in.

Main.S3Tables.create_table — Function

create_table(format, name, namespace, table_bucket_arn)
create_table(format, name, namespace, table_bucket_arn, params::Dict{String,<:Any})

Creates a new table associated with the given namespace in a table bucket. For more information, see Creating an Amazon S3 table in the Amazon Simple Storage Service User Guide.

Permissions

You must have the s3tables:CreateTable permission to use this operation.

If you use this operation with the optional metadata request parameter you must have the s3tables:PutTableData permission.

If you use this operation with the optional encryption_configuration request parameter you must have the s3tables:PutTableEncryption permission.

If you use this operation with the storage_class_configuration request parameter, you must have the s3tables:PutTableStorageClass permission.

To create a table with tags, you must have the s3tables:TagResource permission in addition to s3tables:CreateTable permission.

Note

Additionally, If you choose SSE-KMS encryption you must grant the S3 Tables maintenance principal access to your KMS key. For more information, see Permissions requirements for S3 Tables SSE-KMS encryption.

Arguments

format: The format for the table.

name: The name for the table.

namespace: The namespace to associated with the table.

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket to create the table in.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"encryptionConfiguration": The encryption configuration to use for the table. This configuration specifies the encryption algorithm and, if using SSE-KMS, the KMS key to use for encrypting the table.

Note

If you choose SSE-KMS encryption you must grant the S3 Tables maintenance principal access to your KMS key. For more information, see Permissions requirements for S3 Tables SSE-KMS encryption.

"metadata": The metadata for the table.

"storageClassConfiguration": The storage class configuration for the table. If not specified, the table inherits the storage class configuration from its table bucket. Specify this parameter to override the bucket's default storage class for this table.

"tags": A map of user-defined tags that you would like to apply to the table that you are creating. A tag is a key-value pair that you apply to your resources. Tags can help you organize, track costs for, and control access to resources. For more information, see Tagging for cost allocation or attribute-based access control (ABAC).

Note

You must have the s3tables:TagResource permission in addition to s3tables:CreateTable permission to create a table with tags.

Main.S3Tables.create_table_bucket — Function

create_table_bucket(name)
create_table_bucket(name, params::Dict{String,<:Any})

Creates a table bucket. For more information, see Creating a table bucket in the Amazon Simple Storage Service User Guide.

Permissions

You must have the s3tables:CreateTableBucket permission to use this operation.

If you use this operation with the optional encryptionConfiguration parameter you must have the s3tables:PutTableBucketEncryption permission.

If you use this operation with the storage_class_configuration request parameter, you must have the s3tables:PutTableBucketStorageClass permission.

To create a table bucket with tags, you must have the s3tables:TagResource permission in addition to s3tables:CreateTableBucket permission.

Arguments

name: The name for the table bucket.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"encryptionConfiguration": The encryption configuration to use for the table bucket. This configuration specifies the default encryption settings that will be applied to all tables created in this bucket unless overridden at the table level. The configuration includes the encryption algorithm and, if using SSE-KMS, the KMS key to use.

"storageClassConfiguration": The default storage class configuration for the table bucket. This configuration will be applied to all new tables created in this bucket unless overridden at the table level. If not specified, the service default storage class will be used.

"tags": A map of user-defined tags that you would like to apply to the table bucket that you are creating. A tag is a key-value pair that you apply to your resources. Tags can help you organize and control access to resources. For more information, see Tagging for cost allocation or attribute-based access control (ABAC).

Note

You must have the s3tables:TagResource permission in addition to s3tables:CreateTableBucket permisson to create a table bucket with tags.

Main.S3Tables.delete_namespace — Function

delete_namespace(namespace, table_bucket_arn)
delete_namespace(namespace, table_bucket_arn, params::Dict{String,<:Any})

Deletes a namespace. For more information, see Delete a namespace in the Amazon Simple Storage Service User Guide.

Permissions

You must have the s3tables:DeleteNamespace permission to use this operation.

Arguments

namespace: The name of the namespace.

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket associated with the namespace.

Main.S3Tables.delete_table — Function

delete_table(name, namespace, table_bucket_arn)
delete_table(name, namespace, table_bucket_arn, params::Dict{String,<:Any})

Deletes a table. For more information, see Deleting an Amazon S3 table in the Amazon Simple Storage Service User Guide.

Permissions

You must have the s3tables:DeleteTable permission to use this operation.

Arguments

name: The name of the table.

namespace: The namespace associated with the table.

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket that contains the table.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"versionToken": The version token of the table.

Main.S3Tables.delete_table_bucket — Function

delete_table_bucket(table_bucket_arn)
delete_table_bucket(table_bucket_arn, params::Dict{String,<:Any})

Deletes a table bucket. For more information, see Deleting a table bucket in the Amazon Simple Storage Service User Guide.

Permissions

You must have the s3tables:DeleteTableBucket permission to use this operation.

Arguments

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket.

Main.S3Tables.delete_table_bucket_encryption — Function

delete_table_bucket_encryption(table_bucket_arn)
delete_table_bucket_encryption(table_bucket_arn, params::Dict{String,<:Any})

Deletes the encryption configuration for a table bucket.

Permissions

You must have the s3tables:DeleteTableBucketEncryption permission to use this operation.

Arguments

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket.

Main.S3Tables.delete_table_bucket_metrics_configuration — Function

delete_table_bucket_metrics_configuration(table_bucket_arn)
delete_table_bucket_metrics_configuration(table_bucket_arn, params::Dict{String,<:Any})

Deletes the metrics configuration for a table bucket.

Permissions

You must have the s3tables:DeleteTableBucketMetricsConfiguration permission to use this operation.

Arguments

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket.

Main.S3Tables.delete_table_bucket_policy — Function

delete_table_bucket_policy(table_bucket_arn)
delete_table_bucket_policy(table_bucket_arn, params::Dict{String,<:Any})

Deletes a table bucket policy. For more information, see Deleting a table bucket policy in the Amazon Simple Storage Service User Guide.

Permissions

You must have the s3tables:DeleteTableBucketPolicy permission to use this operation.

Arguments

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket.

Main.S3Tables.delete_table_bucket_replication — Function

delete_table_bucket_replication(table_bucket_arn)
delete_table_bucket_replication(table_bucket_arn, params::Dict{String,<:Any})

Deletes the replication configuration for a table bucket. After deletion, new table updates will no longer be replicated to destination buckets, though existing replicated tables will remain in destination buckets.

Permissions

You must have the s3tables:DeleteTableBucketReplication permission to use this operation.

Arguments

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"versionToken": A version token from a previous GetTableBucketReplication call. Use this token to ensure you're deleting the expected version of the configuration.

Main.S3Tables.delete_table_policy — Function

delete_table_policy(name, namespace, table_bucket_arn)
delete_table_policy(name, namespace, table_bucket_arn, params::Dict{String,<:Any})

Deletes a table policy. For more information, see Deleting a table policy in the Amazon Simple Storage Service User Guide.

Permissions

You must have the s3tables:DeleteTablePolicy permission to use this operation.

Arguments

name: The table name.

namespace: The namespace associated with the table.

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket that contains the table.

Main.S3Tables.delete_table_replication — Function

delete_table_replication(table_arn, version_token)
delete_table_replication(table_arn, version_token, params::Dict{String,<:Any})

Deletes the replication configuration for a specific table. After deletion, new updates to this table will no longer be replicated to destination tables, though existing replicated copies will remain in destination buckets.

Permissions

You must have the s3tables:DeleteTableReplication permission to use this operation.

Arguments

table_arn: The Amazon Resource Name (ARN) of the table.

version_token: A version token from a previous GetTableReplication call. Use this token to ensure you're deleting the expected version of the configuration.

Main.S3Tables.get_namespace — Function

get_namespace(namespace, table_bucket_arn)
get_namespace(namespace, table_bucket_arn, params::Dict{String,<:Any})

Gets details about a namespace. For more information, see Table namespaces in the Amazon Simple Storage Service User Guide.

Permissions

You must have the s3tables:GetNamespace permission to use this operation.

Arguments

namespace: The name of the namespace.

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket.

Main.S3Tables.get_table — Function

get_table()
get_table(params::Dict{String,<:Any})

Gets details about a table. For more information, see S3 Tables in the Amazon Simple Storage Service User Guide.

Permissions

You must have the s3tables:GetTable permission to use this operation.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"name": The name of the table.

"namespace": The name of the namespace the table is associated with.

"tableArn": The Amazon Resource Name (ARN) of the table.

"tableBucketARN": The Amazon Resource Name (ARN) of the table bucket associated with the table.

Main.S3Tables.get_table_bucket — Function

get_table_bucket(table_bucket_arn)
get_table_bucket(table_bucket_arn, params::Dict{String,<:Any})

Gets details on a table bucket. For more information, see Viewing details about an Amazon S3 table bucket in the Amazon Simple Storage Service User Guide.

Permissions

You must have the s3tables:GetTableBucket permission to use this operation.

Arguments

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket.

Main.S3Tables.get_table_bucket_encryption — Function

get_table_bucket_encryption(table_bucket_arn)
get_table_bucket_encryption(table_bucket_arn, params::Dict{String,<:Any})

Gets the encryption configuration for a table bucket.

Permissions

You must have the s3tables:GetTableBucketEncryption permission to use this operation.

Arguments

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket.

Main.S3Tables.get_table_bucket_maintenance_configuration — Function

get_table_bucket_maintenance_configuration(table_bucket_arn)
get_table_bucket_maintenance_configuration(table_bucket_arn, params::Dict{String,<:Any})

Gets details about a maintenance configuration for a given table bucket. For more information, see Amazon S3 table bucket maintenance in the Amazon Simple Storage Service User Guide.

Permissions

You must have the s3tables:GetTableBucketMaintenanceConfiguration permission to use this operation.

Arguments

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket associated with the maintenance configuration.

Main.S3Tables.get_table_bucket_metrics_configuration — Function

get_table_bucket_metrics_configuration(table_bucket_arn)
get_table_bucket_metrics_configuration(table_bucket_arn, params::Dict{String,<:Any})

Gets the metrics configuration for a table bucket.

Permissions

You must have the s3tables:GetTableBucketMetricsConfiguration permission to use this operation.

Arguments

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket.

Main.S3Tables.get_table_bucket_policy — Function

get_table_bucket_policy(table_bucket_arn)
get_table_bucket_policy(table_bucket_arn, params::Dict{String,<:Any})

Gets details about a table bucket policy. For more information, see Viewing a table bucket policy in the Amazon Simple Storage Service User Guide.

Permissions

You must have the s3tables:GetTableBucketPolicy permission to use this operation.

Arguments

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket.

Main.S3Tables.get_table_bucket_replication — Function

get_table_bucket_replication(table_bucket_arn)
get_table_bucket_replication(table_bucket_arn, params::Dict{String,<:Any})

Retrieves the replication configuration for a table bucket.This operation returns the IAM role, versionToken, and replication rules that define how tables in this bucket are replicated to other buckets.

Permissions

You must have the s3tables:GetTableBucketReplication permission to use this operation.

Arguments

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket.

Main.S3Tables.get_table_bucket_storage_class — Function

get_table_bucket_storage_class(table_bucket_arn)
get_table_bucket_storage_class(table_bucket_arn, params::Dict{String,<:Any})

Retrieves the storage class configuration for a specific table. This allows you to view the storage class settings that apply to an individual table, which may differ from the table bucket's default configuration.

Permissions

You must have the s3tables:GetTableBucketStorageClass permission to use this operation.

Arguments

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket.

Main.S3Tables.get_table_encryption — Function

get_table_encryption(name, namespace, table_bucket_arn)
get_table_encryption(name, namespace, table_bucket_arn, params::Dict{String,<:Any})

Gets the encryption configuration for a table.

Permissions

You must have the s3tables:GetTableEncryption permission to use this operation.

Arguments

name: The name of the table.

namespace: The namespace associated with the table.

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket containing the table.

Main.S3Tables.get_table_maintenance_configuration — Function

get_table_maintenance_configuration(name, namespace, table_bucket_arn)
get_table_maintenance_configuration(name, namespace, table_bucket_arn, params::Dict{String,<:Any})

Gets details about the maintenance configuration of a table. For more information, see S3 Tables maintenance in the Amazon Simple Storage Service User Guide.

Permissions

You must have the s3tables:GetTableMaintenanceConfiguration permission to use this operation.

You must have the s3tables:GetTableData permission to use set the compaction strategy to sort or zorder.

Arguments

name: The name of the table.

namespace: The namespace associated with the table.

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket.

Main.S3Tables.get_table_maintenance_job_status — Function

get_table_maintenance_job_status(name, namespace, table_bucket_arn)
get_table_maintenance_job_status(name, namespace, table_bucket_arn, params::Dict{String,<:Any})

Gets the status of a maintenance job for a table. For more information, see S3 Tables maintenance in the Amazon Simple Storage Service User Guide.

Permissions

You must have the s3tables:GetTableMaintenanceJobStatus permission to use this operation.

Arguments

name: The name of the table containing the maintenance job status you want to check.

namespace: The name of the namespace the table is associated with.

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket.

Main.S3Tables.get_table_metadata_location — Function

get_table_metadata_location(name, namespace, table_bucket_arn)
get_table_metadata_location(name, namespace, table_bucket_arn, params::Dict{String,<:Any})

Gets the location of the table metadata.

Permissions

You must have the s3tables:GetTableMetadataLocation permission to use this operation.

Arguments

name: The name of the table.

namespace: The namespace of the table.

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket.

Main.S3Tables.get_table_policy — Function

get_table_policy(name, namespace, table_bucket_arn)
get_table_policy(name, namespace, table_bucket_arn, params::Dict{String,<:Any})

Gets details about a table policy. For more information, see Viewing a table policy in the Amazon Simple Storage Service User Guide.

Permissions

You must have the s3tables:GetTablePolicy permission to use this operation.

Arguments

name: The name of the table.

namespace: The namespace associated with the table.

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket that contains the table.

Main.S3Tables.get_table_record_expiration_configuration — Function

get_table_record_expiration_configuration(table_arn)
get_table_record_expiration_configuration(table_arn, params::Dict{String,<:Any})

Retrieves the expiration configuration settings for records in a table, and the status of the configuration. If the status of the configuration is enabled, records expire and are automatically removed from the table after the specified number of days.

Permissions

You must have the s3tables:GetTableRecordExpirationConfiguration permission to use this operation.

Arguments

table_arn: The Amazon Resource Name (ARN) of the table.

Main.S3Tables.get_table_record_expiration_job_status — Function

get_table_record_expiration_job_status(table_arn)
get_table_record_expiration_job_status(table_arn, params::Dict{String,<:Any})

Retrieves the status, metrics, and details of the latest record expiration job for a table. This includes when the job ran, and whether it succeeded or failed. If the job ran successfully, this also includes statistics about the records that were removed.

Permissions

You must have the s3tables:GetTableRecordExpirationJobStatus permission to use this operation.

Arguments

table_arn: The Amazon Resource Name (ARN) of the table.

Main.S3Tables.get_table_replication — Function

get_table_replication(table_arn)
get_table_replication(table_arn, params::Dict{String,<:Any})

Retrieves the replication configuration for a specific table.

Permissions

You must have the s3tables:GetTableReplication permission to use this operation.

Arguments

table_arn: The Amazon Resource Name (ARN) of the table.

Main.S3Tables.get_table_replication_status — Function

get_table_replication_status(table_arn)
get_table_replication_status(table_arn, params::Dict{String,<:Any})

Retrieves the replication status for a table, including the status of replication to each destination. This operation provides visibility into replication health and progress.

Permissions

You must have the s3tables:GetTableReplicationStatus permission to use this operation.

Arguments

table_arn: The Amazon Resource Name (ARN) of the table.

Main.S3Tables.get_table_storage_class — Function

get_table_storage_class(name, namespace, table_bucket_arn)
get_table_storage_class(name, namespace, table_bucket_arn, params::Dict{String,<:Any})

Retrieves the storage class configuration for a specific table. This allows you to view the storage class settings that apply to an individual table, which may differ from the table bucket's default configuration.

Permissions

You must have the s3tables:GetTableStorageClass permission to use this operation.

Arguments

name: The name of the table.

namespace: The namespace associated with the table.

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket that contains the table.

Main.S3Tables.list_namespaces — Function

list_namespaces(table_bucket_arn)
list_namespaces(table_bucket_arn, params::Dict{String,<:Any})

Lists the namespaces within a table bucket. For more information, see Table namespaces in the Amazon Simple Storage Service User Guide.

Permissions

You must have the s3tables:ListNamespaces permission to use this operation.

Arguments

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"continuationToken": ContinuationToken indicates to Amazon S3 that the list is being continued on this bucket with a token. ContinuationToken is obfuscated and is not a real key. You can use this ContinuationToken for pagination of the list results.

"maxNamespaces": The maximum number of namespaces to return in the list.

"prefix": The prefix of the namespaces.

Main.S3Tables.list_table_buckets — Function

list_table_buckets()
list_table_buckets(params::Dict{String,<:Any})

Lists table buckets for your account. For more information, see S3 Table buckets in the Amazon Simple Storage Service User Guide.

Permissions

You must have the s3tables:ListTableBuckets permission to use this operation.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"continuationToken": ContinuationToken indicates to Amazon S3 that the list is being continued on this bucket with a token. ContinuationToken is obfuscated and is not a real key. You can use this ContinuationToken for pagination of the list results.

"maxBuckets": The maximum number of table buckets to return in the list.

"prefix": The prefix of the table buckets.

"type": The type of table buckets to filter by in the list.

Main.S3Tables.list_tables — Function

list_tables(table_bucket_arn)
list_tables(table_bucket_arn, params::Dict{String,<:Any})

List tables in the given table bucket. For more information, see S3 Tables in the Amazon Simple Storage Service User Guide.

Permissions

You must have the s3tables:ListTables permission to use this operation.

Arguments

table_bucket_arn: The Amazon resource Name (ARN) of the table bucket.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"continuationToken": ContinuationToken indicates to Amazon S3 that the list is being continued on this bucket with a token. ContinuationToken is obfuscated and is not a real key. You can use this ContinuationToken for pagination of the list results.

"maxTables": The maximum number of tables to return.

"namespace": The namespace of the tables.

"prefix": The prefix of the tables.

Main.S3Tables.list_tags_for_resource — Function

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists all of the tags applied to a specified Amazon S3 Tables resource. Each tag is a label consisting of a key and value pair. Tags can help you organize, track costs for, and control access to resources.

Note

For a list of S3 resources that support tagging, see Managing tags for Amazon S3 resources.

Permissions

For tables and table buckets, you must have the s3tables:ListTagsForResource permission to use this operation.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the Amazon S3 Tables resource that you want to list tags for. The tagged resource can be a table bucket or a table. For a list of all S3 resources that support tagging, see Managing tags for Amazon S3 resources.

Main.S3Tables.put_table_bucket_encryption — Function

put_table_bucket_encryption(encryption_configuration, table_bucket_arn)
put_table_bucket_encryption(encryption_configuration, table_bucket_arn, params::Dict{String,<:Any})

Sets the encryption configuration for a table bucket.

Permissions

You must have the s3tables:PutTableBucketEncryption permission to use this operation.

Note

If you choose SSE-KMS encryption you must grant the S3 Tables maintenance principal access to your KMS key. For more information, see Permissions requirements for S3 Tables SSE-KMS encryption in the Amazon Simple Storage Service User Guide.

Arguments

encryption_configuration: The encryption configuration to apply to the table bucket.

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket.

Main.S3Tables.put_table_bucket_maintenance_configuration — Function

put_table_bucket_maintenance_configuration(table_bucket_arn, type, value)
put_table_bucket_maintenance_configuration(table_bucket_arn, type, value, params::Dict{String,<:Any})

Creates a new maintenance configuration or replaces an existing maintenance configuration for a table bucket. For more information, see Amazon S3 table bucket maintenance in the Amazon Simple Storage Service User Guide.

Permissions

You must have the s3tables:PutTableBucketMaintenanceConfiguration permission to use this operation.

Arguments

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket associated with the maintenance configuration.

type: The type of the maintenance configuration.

value: Defines the values of the maintenance configuration for the table bucket.

Main.S3Tables.put_table_bucket_metrics_configuration — Function

put_table_bucket_metrics_configuration(table_bucket_arn)
put_table_bucket_metrics_configuration(table_bucket_arn, params::Dict{String,<:Any})

Sets the metrics configuration for a table bucket.

Permissions

You must have the s3tables:PutTableBucketMetricsConfiguration permission to use this operation.

Arguments

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket.

Main.S3Tables.put_table_bucket_policy — Function

put_table_bucket_policy(resource_policy, table_bucket_arn)
put_table_bucket_policy(resource_policy, table_bucket_arn, params::Dict{String,<:Any})

Creates a new table bucket policy or replaces an existing table bucket policy for a table bucket. For more information, see Adding a table bucket policy in the Amazon Simple Storage Service User Guide.

Permissions

You must have the s3tables:PutTableBucketPolicy permission to use this operation.

Arguments

resource_policy: The JSON that defines the policy.

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket.

Main.S3Tables.put_table_bucket_replication — Function

put_table_bucket_replication(configuration, table_bucket_arn)
put_table_bucket_replication(configuration, table_bucket_arn, params::Dict{String,<:Any})

Creates or updates the replication configuration for a table bucket. This operation defines how tables in the source bucket are replicated to destination buckets. Replication helps ensure data availability and disaster recovery across regions or accounts.

Permissions

You must have the s3tables:PutTableBucketReplication permission to use this operation. The IAM role specified in the configuration must have permissions to read from the source bucket and write permissions to all destination buckets.

You must also have the following permissions:

s3tables:GetTable permission on the source table.

s3tables:ListTables permission on the bucket containing the table.

s3tables:CreateTable permission for the destination.

s3tables:CreateNamespace permission for the destination.

s3tables:GetTableMaintenanceConfig permission for the source bucket.

s3tables:PutTableMaintenanceConfig permission for the destination bucket.

You must have iam:PassRole permission with condition allowing roles to be passed to replication.s3tables.amazonaws.com.

Arguments

configuration: The replication configuration to apply, including the IAM role and replication rules.

table_bucket_arn: The Amazon Resource Name (ARN) of the source table bucket.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"versionToken": A version token from a previous GetTableBucketReplication call. Use this token to ensure you're updating the expected version of the configuration.

Main.S3Tables.put_table_bucket_storage_class — Function

put_table_bucket_storage_class(storage_class_configuration, table_bucket_arn)
put_table_bucket_storage_class(storage_class_configuration, table_bucket_arn, params::Dict{String,<:Any})

Sets or updates the storage class configuration for a table bucket. This configuration serves as the default storage class for all new tables created in the bucket, allowing you to optimize storage costs at the bucket level.

Permissions

You must have the s3tables:PutTableBucketStorageClass permission to use this operation.

Arguments

storage_class_configuration: The storage class configuration to apply to the table bucket. This configuration will serve as the default for new tables created in this bucket.

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket.

Main.S3Tables.put_table_maintenance_configuration — Function

put_table_maintenance_configuration(name, namespace, table_bucket_arn, type, value)
put_table_maintenance_configuration(name, namespace, table_bucket_arn, type, value, params::Dict{String,<:Any})

Creates a new maintenance configuration or replaces an existing maintenance configuration for a table. For more information, see S3 Tables maintenance in the Amazon Simple Storage Service User Guide.

Permissions

You must have the s3tables:PutTableMaintenanceConfiguration permission to use this operation.

Arguments

name: The name of the table.

namespace: The namespace of the table.

table_bucket_arn: The Amazon Resource Name (ARN) of the table associated with the maintenance configuration.

type: The type of the maintenance configuration.

value: Defines the values of the maintenance configuration for the table.

Main.S3Tables.put_table_policy — Function

put_table_policy(name, namespace, resource_policy, table_bucket_arn)
put_table_policy(name, namespace, resource_policy, table_bucket_arn, params::Dict{String,<:Any})

Creates a new table policy or replaces an existing table policy for a table. For more information, see Adding a table policy in the Amazon Simple Storage Service User Guide.

Permissions

You must have the s3tables:PutTablePolicy permission to use this operation.

Arguments

name: The name of the table.

namespace: The namespace associated with the table.

resource_policy: The JSON that defines the policy.

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket that contains the table.

Main.S3Tables.put_table_record_expiration_configuration — Function

put_table_record_expiration_configuration(table_arn, value)
put_table_record_expiration_configuration(table_arn, value, params::Dict{String,<:Any})

Creates or updates the expiration configuration settings for records in a table, including the status of the configuration. If you enable record expiration for a table, records expire and are automatically removed from the table after the number of days that you specify.

Permissions

You must have the s3tables:PutTableRecordExpirationConfiguration permission to use this operation.

Arguments

table_arn: The Amazon Resource Name (ARN) of the table.

value: The record expiration configuration to apply to the table, including the status (enabled or disabled) and retention period in days.

Main.S3Tables.put_table_replication — Function

put_table_replication(configuration, table_arn)
put_table_replication(configuration, table_arn, params::Dict{String,<:Any})

Creates or updates the replication configuration for a specific table. This operation allows you to define table-level replication independently of bucket-level replication, providing granular control over which tables are replicated and where.

Permissions

You must have the s3tables:PutTableReplication permission to use this operation. The IAM role specified in the configuration must have permissions to read from the source table and write to all destination tables.

You must also have the following permissions:

s3tables:GetTable permission on the source table being replicated.

s3tables:CreateTable permission for the destination.

s3tables:CreateNamespace permission for the destination.

s3tables:GetTableMaintenanceConfig permission for the source table.

s3tables:PutTableMaintenanceConfig permission for the destination table.

You must have iam:PassRole permission with condition allowing roles to be passed to replication.s3tables.amazonaws.com.

Arguments

configuration: The replication configuration to apply to the table, including the IAM role and replication rules.

table_arn: The Amazon Resource Name (ARN) of the source table.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"versionToken": A version token from a previous GetTableReplication call. Use this token to ensure you're updating the expected version of the configuration.

Main.S3Tables.rename_table — Function

rename_table(name, namespace, table_bucket_arn)
rename_table(name, namespace, table_bucket_arn, params::Dict{String,<:Any})

Renames a table or a namespace. For more information, see S3 Tables in the Amazon Simple Storage Service User Guide.

Permissions

You must have the s3tables:RenameTable permission to use this operation.

Arguments

name: The current name of the table.

namespace: The namespace associated with the table.

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"newName": The new name for the table.

"newNamespaceName": The new name for the namespace.

"versionToken": The version token of the table.

Main.S3Tables.tag_resource — Function

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Applies one or more user-defined tags to an Amazon S3 Tables resource or updates existing tags. Each tag is a label consisting of a key and value pair. Tags can help you organize, track costs for, and control access to your resources. You can add up to 50 tags for each S3 resource.

Note

For a list of S3 resources that support tagging, see Managing tags for Amazon S3 resources.

Permissions

For tables and table buckets, you must have the s3tables:TagResource permission to use this operation.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the Amazon S3 Tables resource that you're applying tags to. The tagged resource can be a table bucket or a table. For a list of all S3 resources that support tagging, see Managing tags for Amazon S3 resources.

tags: The user-defined tag that you want to add to the specified S3 Tables resource. For more information, see Tagging for cost allocation or attribute-based access control (ABAC).

Main.S3Tables.untag_resource — Function

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes the specified user-defined tags from an Amazon S3 Tables resource. You can pass one or more tag keys.

Note

For a list of S3 resources that support tagging, see Managing tags for Amazon S3 resources.

Permissions

For tables and table buckets, you must have the s3tables:UntagResource permission to use this operation.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the Amazon S3 Tables resource that you're removing tags from. The tagged resource can be a table bucket or a table. For a list of all S3 resources that support tagging, see Managing tags for Amazon S3 resources.

tag_keys: The array of tag keys that you're removing from the S3 Tables resource. For more information, see Tagging for cost allocation or attribute-based access control (ABAC).

Main.S3Tables.update_table_metadata_location — Function

update_table_metadata_location(metadata_location, name, namespace, table_bucket_arn, version_token)
update_table_metadata_location(metadata_location, name, namespace, table_bucket_arn, version_token, params::Dict{String,<:Any})

Updates the metadata location for a table. The metadata location of a table must be an S3 URI that begins with the table's warehouse location. The metadata location for an Apache Iceberg table must end with .metadata.json, or if the metadata file is Gzip-compressed, .metadata.json.gz.

Permissions

You must have the s3tables:UpdateTableMetadataLocation permission to use this operation.

Arguments

metadata_location: The new metadata location for the table.

name: The name of the table.

namespace: The namespace of the table.

table_bucket_arn: The Amazon Resource Name (ARN) of the table bucket.

version_token: The version token of the table.