S3Vectors

This page documents function available when using the S3Vectors module, created with @service S3Vectors.

Index

Main.S3Vectors.create_index

Main.S3Vectors.create_vector_bucket

Main.S3Vectors.delete_index

Main.S3Vectors.delete_vector_bucket

Main.S3Vectors.delete_vector_bucket_policy

Main.S3Vectors.delete_vectors

Main.S3Vectors.get_index

Main.S3Vectors.get_vector_bucket

Main.S3Vectors.get_vector_bucket_policy

Main.S3Vectors.get_vectors

Main.S3Vectors.list_indexes

Main.S3Vectors.list_tags_for_resource

Main.S3Vectors.list_vector_buckets

Main.S3Vectors.list_vectors

Main.S3Vectors.put_vector_bucket_policy

Main.S3Vectors.put_vectors

Main.S3Vectors.query_vectors

Main.S3Vectors.tag_resource

Main.S3Vectors.untag_resource

Documentation

Main.S3Vectors.create_index — Function

create_index(data_type, dimension, distance_metric, index_name)
create_index(data_type, dimension, distance_metric, index_name, params::Dict{String,<:Any})

Creates a vector index within a vector bucket. To specify the vector bucket, you must use either the vector bucket name or the vector bucket Amazon Resource Name (ARN).

Permissions

You must have the s3vectors:CreateIndex permission to use this operation.

You must have the s3vectors:TagResource permission in addition to s3vectors:CreateIndex permission to create a vector index with tags.

Arguments

data_type: The data type of the vectors to be inserted into the vector index.

dimension: The dimensions of the vectors to be inserted into the vector index.

distance_metric: The distance metric to be used for similarity search.

index_name: The name of the vector index to create.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"encryptionConfiguration": The encryption configuration for a vector index. By default, if you don't specify, all new vectors in the vector index will use the encryption configuration of the vector bucket.

"metadataConfiguration": The metadata configuration for the vector index.

"tags": An array of user-defined tags that you would like to apply to the vector index that you are creating. A tag is a key-value pair that you apply to your resources. Tags can help you organize, track costs, and control access to resources. For more information, see Tagging for cost allocation or attribute-based access control (ABAC).

Note

You must have the s3vectors:TagResource permission in addition to s3vectors:CreateIndex permission to create a vector index with tags.

"vectorBucketArn": The Amazon Resource Name (ARN) of the vector bucket to create the vector index in.

"vectorBucketName": The name of the vector bucket to create the vector index in.

Main.S3Vectors.create_vector_bucket — Function

create_vector_bucket(vector_bucket_name)
create_vector_bucket(vector_bucket_name, params::Dict{String,<:Any})

Creates a vector bucket in the Amazon Web Services Region that you want your bucket to be in.

Permissions

You must have the s3vectors:CreateVectorBucket permission to use this operation.

You must have the s3vectors:TagResource permission in addition to s3vectors:CreateVectorBucket permission to create a vector bucket with tags.

Arguments

vector_bucket_name: The name of the vector bucket to create.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"encryptionConfiguration": The encryption configuration for the vector bucket. By default, if you don't specify, all new vectors in Amazon S3 vector buckets use server-side encryption with Amazon S3 managed keys (SSE-S3), specifically AES256.

"tags": An array of user-defined tags that you would like to apply to the vector bucket that you are creating. A tag is a key-value pair that you apply to your resources. Tags can help you organize and control access to resources. For more information, see Tagging for cost allocation or attribute-based access control (ABAC).

Note

You must have the s3vectors:TagResource permission in addition to s3vectors:CreateVectorBucket permission to create a vector bucket with tags.

Main.S3Vectors.delete_index — Function

delete_index()
delete_index(params::Dict{String,<:Any})

Deletes a vector index. To specify the vector index, you can either use both the vector bucket name and vector index name, or use the vector index Amazon Resource Name (ARN).

Permissions

You must have the s3vectors:DeleteIndex permission to use this operation.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"indexArn": The ARN of the vector index to delete.

"indexName": The name of the vector index to delete.

"vectorBucketName": The name of the vector bucket that contains the vector index.

Main.S3Vectors.delete_vector_bucket — Function

delete_vector_bucket()
delete_vector_bucket(params::Dict{String,<:Any})

Deletes a vector bucket. All vector indexes in the vector bucket must be deleted before the vector bucket can be deleted. To perform this operation, you must use either the vector bucket name or the vector bucket Amazon Resource Name (ARN).

Permissions

You must have the s3vectors:DeleteVectorBucket permission to use this operation.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"vectorBucketArn": The ARN of the vector bucket to delete.

"vectorBucketName": The name of the vector bucket to delete.

Main.S3Vectors.delete_vector_bucket_policy — Function

delete_vector_bucket_policy()
delete_vector_bucket_policy(params::Dict{String,<:Any})

Deletes a vector bucket policy. To specify the bucket, you must use either the vector bucket name or the vector bucket Amazon Resource Name (ARN).

Permissions

You must have the s3vectors:DeleteVectorBucketPolicy permission to use this operation.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"vectorBucketArn": The ARN of the vector bucket to delete the policy from.

"vectorBucketName": The name of the vector bucket to delete the policy from.

Main.S3Vectors.delete_vectors — Function

delete_vectors(keys)
delete_vectors(keys, params::Dict{String,<:Any})

Deletes one or more vectors in a vector index. To specify the vector index, you can either use both the vector bucket name and vector index name, or use the vector index Amazon Resource Name (ARN).

Permissions

You must have the s3vectors:DeleteVectors permission to use this operation.

Arguments

keys: The keys of the vectors to delete.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"indexArn": The ARN of the vector index that contains a vector you want to delete.

"indexName": The name of the vector index that contains a vector you want to delete.

"vectorBucketName": The name of the vector bucket that contains the vector index.

Main.S3Vectors.get_index — Function

get_index()
get_index(params::Dict{String,<:Any})

Returns vector index attributes. To specify the vector index, you can either use both the vector bucket name and the vector index name, or use the vector index Amazon Resource Name (ARN).

Permissions

You must have the s3vectors:GetIndex permission to use this operation.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"indexArn": The ARN of the vector index.

"indexName": The name of the vector index.

"vectorBucketName": The name of the vector bucket that contains the vector index.

Main.S3Vectors.get_vector_bucket — Function

get_vector_bucket()
get_vector_bucket(params::Dict{String,<:Any})

Returns vector bucket attributes. To specify the bucket, you must use either the vector bucket name or the vector bucket Amazon Resource Name (ARN).

Permissions

You must have the s3vectors:GetVectorBucket permission to use this operation.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"vectorBucketArn": The ARN of the vector bucket to retrieve information about.

"vectorBucketName": The name of the vector bucket to retrieve information about.

Main.S3Vectors.get_vector_bucket_policy — Function

get_vector_bucket_policy()
get_vector_bucket_policy(params::Dict{String,<:Any})

Gets details about a vector bucket policy. To specify the bucket, you must use either the vector bucket name or the vector bucket Amazon Resource Name (ARN).

Permissions

You must have the s3vectors:GetVectorBucketPolicy permission to use this operation.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"vectorBucketArn": The ARN of the vector bucket.

"vectorBucketName": The name of the vector bucket.

Main.S3Vectors.get_vectors — Function

get_vectors(keys)
get_vectors(keys, params::Dict{String,<:Any})

Returns vector attributes. To specify the vector index, you can either use both the vector bucket name and the vector index name, or use the vector index Amazon Resource Name (ARN).

Permissions

You must have the s3vectors:GetVectors permission to use this operation.

Arguments

keys: The names of the vectors you want to return attributes for.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"indexArn": The ARN of the vector index.

"indexName": The name of the vector index.

"returnData": Indicates whether to include the vector data in the response. The default value is false.

"returnMetadata": Indicates whether to include metadata in the response. The default value is false.

"vectorBucketName": The name of the vector bucket that contains the vector index.

Main.S3Vectors.list_indexes — Function

list_indexes()
list_indexes(params::Dict{String,<:Any})

Returns a list of all the vector indexes within the specified vector bucket. To specify the bucket, you must use either the vector bucket name or the vector bucket Amazon Resource Name (ARN).

Permissions

You must have the s3vectors:ListIndexes permission to use this operation.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of items to be returned in the response.

"nextToken": The previous pagination token.

"prefix": Limits the response to vector indexes that begin with the specified prefix.

"vectorBucketArn": The ARN of the vector bucket that contains the vector indexes.

"vectorBucketName": The name of the vector bucket that contains the vector indexes.

Main.S3Vectors.list_tags_for_resource — Function

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists all of the tags applied to a specified Amazon S3 Vectors resource. Each tag is a label consisting of a key and value pair. Tags can help you organize, track costs for, and control access to resources.

Note

For a list of S3 resources that support tagging, see Managing tags for Amazon S3 resources.

Permissions

For vector buckets and vector indexes, you must have the s3vectors:ListTagsForResource permission to use this operation.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the Amazon S3 Vectors resource that you want to list tags for. The tagged resource can be a vector bucket or a vector index.

Main.S3Vectors.list_vector_buckets — Function

list_vector_buckets()
list_vector_buckets(params::Dict{String,<:Any})

Returns a list of all the vector buckets that are owned by the authenticated sender of the request.

Permissions

You must have the s3vectors:ListVectorBuckets permission to use this operation.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of vector buckets to be returned in the response.

"nextToken": The previous pagination token.

"prefix": Limits the response to vector buckets that begin with the specified prefix.

Main.S3Vectors.list_vectors — Function

list_vectors()
list_vectors(params::Dict{String,<:Any})

List vectors in the specified vector index. To specify the vector index, you can either use both the vector bucket name and the vector index name, or use the vector index Amazon Resource Name (ARN).

list_vectors operations proceed sequentially; however, for faster performance on a large number of vectors in a vector index, applications can request a parallel list_vectors operation by providing the segmentCount and segmentIndex parameters.

Permissions

You must have the s3vectors:ListVectors permission to use this operation. Additional permissions are required based on the request parameters you specify:

With only s3vectors:ListVectors permission, you can list vector keys when returnData and returnMetadata are both set to false or not specified..

If you set returnData or returnMetadata to true, you must have both s3vectors:ListVectors and s3vectors:GetVectors permissions. The request fails with a 403 Forbidden error if you request vector data or metadata without the s3vectors:GetVectors permission.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"indexArn": The Amazon resource Name (ARN) of the vector index.

"indexName": The name of the vector index.

"maxResults": The maximum number of vectors to return on a page.

If you don't specify maxResults, the ListVectors operation uses a default value of

If the processed dataset size exceeds 1 MB before reaching the maxResults value, the operation stops and returns the vectors that are retrieved up to that point, along with a nextToken that you can use in a subsequent request to retrieve the next set of results.

"nextToken": Pagination token from a previous request. The value of this field is empty for an initial request.

"returnData": If true, the vector data of each vector will be included in the response. The default value is false.

"returnMetadata": If true, the metadata associated with each vector will be included in the response. The default value is false.

"segmentCount": For a parallel ListVectors request, segmentCount represents the total number of vector segments into which the ListVectors operation will be divided. The value of segmentCount corresponds to the number of application workers that will perform the parallel ListVectors operation. For example, if you want to use four application threads to list vectors in a vector index, specify a segmentCount value of

If you specify a segmentCount value of 1, the ListVectors operation will be sequential rather than parallel.

If you specify segmentCount, you must also specify segmentIndex.

"segmentIndex": For a parallel ListVectors request, segmentIndex is the index of the segment from which to list vectors in the current request. It identifies an individual segment to be listed by an application worker.

Segment IDs are zero-based, so the first segment is always 0. For example, if you want to use four application threads to list vectors in a vector index, then the first thread specifies a segmentIndex value of 0, the second thread specifies 1, and so on.

The value of segmentIndex must be less than the value provided for segmentCount.

If you provide segmentIndex, you must also provide segmentCount.

"vectorBucketName": The name of the vector bucket.

Main.S3Vectors.put_vector_bucket_policy — Function

put_vector_bucket_policy(policy)
put_vector_bucket_policy(policy, params::Dict{String,<:Any})

Creates a bucket policy for a vector bucket. To specify the bucket, you must use either the vector bucket name or the vector bucket Amazon Resource Name (ARN).

Permissions

You must have the s3vectors:PutVectorBucketPolicy permission to use this operation.

Arguments

policy: The JSON that defines the policy. For more information about bucket policies for S3 Vectors, see Managing vector bucket policies in the Amazon S3 User Guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"vectorBucketArn": The Amazon Resource Name (ARN) of the vector bucket.

"vectorBucketName": The name of the vector bucket.

Main.S3Vectors.put_vectors — Function

put_vectors(vectors)
put_vectors(vectors, params::Dict{String,<:Any})

Adds one or more vectors to a vector index. To specify the vector index, you can either use both the vector bucket name and the vector index name, or use the vector index Amazon Resource Name (ARN).

For more information about limits, see Limitations and restrictions in the Amazon S3 User Guide.

Note

When inserting vector data into your vector index, you must provide the vector data as float32 (32-bit floating point) values. If you pass higher-precision values to an Amazon Web Services SDK, S3 Vectors converts the values to 32-bit floating point before storing them, and GetVectors, ListVectors, and query_vectors operations return the float32 values. Different Amazon Web Services SDKs may have different default numeric types, so ensure your vectors are properly formatted as float32 values regardless of which SDK you're using. For example, in Python, use numpy.float32 or explicitly cast your values.

Permissions

You must have the s3vectors:PutVectors permission to use this operation.

Arguments

vectors: The vectors to add to a vector index. The number of vectors in a single request must not exceed the resource capacity, otherwise the request will be rejected with the error ServiceUnavailableException with the error message "Currently unable to handle the request".

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"indexArn": The ARN of the vector index where you want to write vectors.

"indexName": The name of the vector index where you want to write vectors.

"vectorBucketName": The name of the vector bucket that contains the vector index.

Main.S3Vectors.query_vectors — Function

query_vectors(query_vector, top_k)
query_vectors(query_vector, top_k, params::Dict{String,<:Any})

Performs an approximate nearest neighbor search query in a vector index using a query vector. By default, it returns the keys of approximate nearest neighbors. You can optionally include the computed distance (between the query vector and each vector in the response), the vector data, and metadata of each vector in the response.

To specify the vector index, you can either use both the vector bucket name and the vector index name, or use the vector index Amazon Resource Name (ARN).

Permissions

You must have the s3vectors:QueryVectors permission to use this operation. Additional permissions are required based on the request parameters you specify:

With only s3vectors:QueryVectors permission, you can retrieve vector keys of approximate nearest neighbors and computed distances between these vectors. This permission is sufficient only when you don't set any metadata filters and don't request vector data or metadata (by keeping the returnMetadata parameter set to false or not specified).

If you specify a metadata filter or set returnMetadata to true, you must have both s3vectors:QueryVectors and s3vectors:GetVectors permissions. The request fails with a 403 Forbidden error if you request metadata filtering, vector data, or metadata without the s3vectors:GetVectors permission.

Arguments

query_vector: The query vector. Ensure that the query vector has the same dimension as the dimension of the vector index that's being queried. For example, if your vector index contains vectors with 384 dimensions, your query vector must also have 384 dimensions.

top_k: The number of results to return for each query.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filter": Metadata filter to apply during the query. For more information about metadata keys, see Metadata filtering in the Amazon S3 User Guide.

"indexArn": The ARN of the vector index that you want to query.

"indexName": The name of the vector index that you want to query.

"returnDistance": Indicates whether to include the computed distance in the response. The default value is false.

"returnMetadata": Indicates whether to include metadata in the response. The default value is false.

"vectorBucketName": The name of the vector bucket that contains the vector index.

Main.S3Vectors.tag_resource — Function

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Applies one or more user-defined tags to an Amazon S3 Vectors resource or updates existing tags. Each tag is a label consisting of a key and value pair. Tags can help you organize, track costs for, and control access to your resources. You can add up to 50 tags for each resource.

Note

For a list of S3 resources that support tagging, see Managing tags for Amazon S3 resources.

Permissions

For vector buckets and vector indexes, you must have the s3vectors:TagResource permission to use this operation.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the Amazon S3 Vectors resource that you're applying tags to. The tagged resource can be a vector bucket or a vector index.

tags: The user-defined tag that you want to add to the specified S3 Vectors resource. For more information, see Tagging for cost allocation or attribute-based access control (ABAC).

Main.S3Vectors.untag_resource — Function

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes the specified user-defined tags from an Amazon S3 Vectors resource. You can pass one or more tag keys.

Note

For a list of S3 resources that support tagging, see Managing tags for Amazon S3 resources.

Permissions

For vector buckets and vector indexes, you must have the s3vectors:UntagResource permission to use this operation.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the Amazon S3 Vectors resource that you're removing tags from. The tagged resource can be a vector bucket or a vector index.

tag_keys: The array of tag keys that you're removing from the S3 Vectors resource. For more information, see Tagging for cost allocation or attribute-based access control (ABAC).