Sagemaker A2I Runtime

This page documents function available when using the Sagemaker_A2I_Runtime module, created with @service Sagemaker_A2I_Runtime.

Index

Main.Sagemaker_A2I_Runtime.delete_human_loop

Main.Sagemaker_A2I_Runtime.describe_human_loop

Main.Sagemaker_A2I_Runtime.list_human_loops

Main.Sagemaker_A2I_Runtime.start_human_loop

Main.Sagemaker_A2I_Runtime.stop_human_loop

Documentation

Main.Sagemaker_A2I_Runtime.delete_human_loop — Method

delete_human_loop(human_loop_name)
delete_human_loop(human_loop_name, params::Dict{String,<:Any})

Deletes the specified human loop for a flow definition. If the human loop was deleted, this operation will return a ResourceNotFoundException.

Arguments

human_loop_name: The name of the human loop that you want to delete.

Main.Sagemaker_A2I_Runtime.describe_human_loop — Method

describe_human_loop(human_loop_name)
describe_human_loop(human_loop_name, params::Dict{String,<:Any})

Returns information about the specified human loop. If the human loop was deleted, this operation will return a ResourceNotFoundException error.

Arguments

human_loop_name: The name of the human loop that you want information about.

Main.Sagemaker_A2I_Runtime.list_human_loops — Method

list_human_loops(flow_definition_arn)
list_human_loops(flow_definition_arn, params::Dict{String,<:Any})

Returns information about human loops, given the specified parameters. If a human loop was deleted, it will not be included.

Arguments

flow_definition_arn: The Amazon Resource Name (ARN) of a flow definition.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"CreationTimeAfter": (Optional) The timestamp of the date when you want the human loops to begin in ISO 8601 format. For example, 2020-02-24.

"CreationTimeBefore": (Optional) The timestamp of the date before which you want the human loops to begin in ISO 8601 format. For example, 2020-02-24.

"MaxResults": The total number of items to return. If the total number of available items is more than the value specified in MaxResults, then a NextToken is returned in the output. You can use this token to display the next page of results.

"NextToken": A token to display the next page of results.

"SortOrder": Optional. The order for displaying results. Valid values: Ascending and Descending.

Main.Sagemaker_A2I_Runtime.start_human_loop — Method

start_human_loop(flow_definition_arn, human_loop_input, human_loop_name)
start_human_loop(flow_definition_arn, human_loop_input, human_loop_name, params::Dict{String,<:Any})

Starts a human loop, provided that at least one activation condition is met.

Arguments

flow_definition_arn: The Amazon Resource Name (ARN) of the flow definition associated with this human loop.

human_loop_input: An object that contains information about the human loop.

human_loop_name: The name of the human loop.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DataAttributes": Attributes of the specified data. Use DataAttributes to specify if your data is free of personally identifiable information and/or free of adult content.

Main.Sagemaker_A2I_Runtime.stop_human_loop — Method

stop_human_loop(human_loop_name)
stop_human_loop(human_loop_name, params::Dict{String,<:Any})

Stops the specified human loop.

Arguments

human_loop_name: The name of the human loop that you want to stop.