Sagemaker Geospatial

This page documents function available when using the Sagemaker_Geospatial module, created with @service Sagemaker_Geospatial.

Index

Main.Sagemaker_Geospatial.delete_earth_observation_job

Main.Sagemaker_Geospatial.delete_vector_enrichment_job

Main.Sagemaker_Geospatial.export_earth_observation_job

Main.Sagemaker_Geospatial.export_vector_enrichment_job

Main.Sagemaker_Geospatial.get_earth_observation_job

Main.Sagemaker_Geospatial.get_raster_data_collection

Main.Sagemaker_Geospatial.get_tile

Main.Sagemaker_Geospatial.get_vector_enrichment_job

Main.Sagemaker_Geospatial.list_earth_observation_jobs

Main.Sagemaker_Geospatial.list_raster_data_collections

Main.Sagemaker_Geospatial.list_tags_for_resource

Main.Sagemaker_Geospatial.list_vector_enrichment_jobs

Main.Sagemaker_Geospatial.search_raster_data_collection

Main.Sagemaker_Geospatial.start_earth_observation_job

Main.Sagemaker_Geospatial.start_vector_enrichment_job

Main.Sagemaker_Geospatial.stop_earth_observation_job

Main.Sagemaker_Geospatial.stop_vector_enrichment_job

Main.Sagemaker_Geospatial.tag_resource

Main.Sagemaker_Geospatial.untag_resource

Documentation

Main.Sagemaker_Geospatial.delete_earth_observation_job — Method

delete_earth_observation_job(arn)
delete_earth_observation_job(arn, params::Dict{String,<:Any})

Use this operation to delete an Earth Observation job.

Arguments

arn: The Amazon Resource Name (ARN) of the Earth Observation job being deleted.

Main.Sagemaker_Geospatial.delete_vector_enrichment_job — Method

delete_vector_enrichment_job(arn)
delete_vector_enrichment_job(arn, params::Dict{String,<:Any})

Use this operation to delete a Vector Enrichment job.

Arguments

arn: The Amazon Resource Name (ARN) of the Vector Enrichment job being deleted.

Main.Sagemaker_Geospatial.export_earth_observation_job — Method

export_earth_observation_job(arn, execution_role_arn, output_config)
export_earth_observation_job(arn, execution_role_arn, output_config, params::Dict{String,<:Any})

Use this operation to export results of an Earth Observation job and optionally source images used as input to the EOJ to an S3 location.

Arguments

arn: The input Amazon Resource Name (ARN) of the Earth Observation job being exported.

execution_role_arn: The Amazon Resource Name (ARN) of the IAM role that you specified for the job.

output_config: An object containing information about the output file.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ExportSourceImages": The source images provided to the Earth Observation job being exported.

Main.Sagemaker_Geospatial.export_vector_enrichment_job — Method

export_vector_enrichment_job(arn, execution_role_arn, output_config)
export_vector_enrichment_job(arn, execution_role_arn, output_config, params::Dict{String,<:Any})

Use this operation to copy results of a Vector Enrichment job to an S3 location.

Arguments

arn: The Amazon Resource Name (ARN) of the Vector Enrichment job.

execution_role_arn: The Amazon Resource Name (ARN) of the IAM rolewith permission to upload to the location in OutputConfig.

output_config: Output location information for exporting Vector Enrichment Job results.

Main.Sagemaker_Geospatial.get_earth_observation_job — Method

get_earth_observation_job(arn)
get_earth_observation_job(arn, params::Dict{String,<:Any})

Get the details for a previously initiated Earth Observation job.

Arguments

arn: The Amazon Resource Name (ARN) of the Earth Observation job.

Main.Sagemaker_Geospatial.get_raster_data_collection — Method

get_raster_data_collection(arn)
get_raster_data_collection(arn, params::Dict{String,<:Any})

Use this operation to get details of a specific raster data collection.

Arguments

arn: The Amazon Resource Name (ARN) of the raster data collection.

Main.Sagemaker_Geospatial.get_tile — Method

get_tile(arn, image_assets, target, x, y, z)
get_tile(arn, image_assets, target, x, y, z, params::Dict{String,<:Any})

Gets a web mercator tile for the given Earth Observation job.

Arguments

arn: The Amazon Resource Name (ARN) of the tile operation.

image_assets: The particular assets or bands to tile.

target: Determines what part of the Earth Observation job to tile. 'INPUT' or 'OUTPUT' are the valid options.

x: The x coordinate of the tile input.

y: The y coordinate of the tile input.

z: The z coordinate of the tile input.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ImageMask": Determines whether or not to return a valid data mask.

"OutputDataType": The output data type of the tile operation.

"OutputFormat": The data format of the output tile. The formats include .npy, .png and .jpg.

"PropertyFilters": Property filters for the imagery to tile.

"TimeRangeFilter": Time range filter applied to imagery to find the images to tile.

Main.Sagemaker_Geospatial.get_vector_enrichment_job — Method

get_vector_enrichment_job(arn)
get_vector_enrichment_job(arn, params::Dict{String,<:Any})

Retrieves details of a Vector Enrichment Job for a given job Amazon Resource Name (ARN).

Arguments

arn: The Amazon Resource Name (ARN) of the Vector Enrichment job.

Main.Sagemaker_Geospatial.list_earth_observation_jobs — Method

list_earth_observation_jobs()
list_earth_observation_jobs(params::Dict{String,<:Any})

Use this operation to get a list of the Earth Observation jobs associated with the calling Amazon Web Services account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The total number of items to return.

"NextToken": If the previous response was truncated, you receive this token. Use it in your next request to receive the next set of results.

"SortBy": The parameter by which to sort the results.

"SortOrder": An optional value that specifies whether you want the results sorted in Ascending or Descending order.

"StatusEquals": A filter that retrieves only jobs with a specific status.

Main.Sagemaker_Geospatial.list_raster_data_collections — Method

list_raster_data_collections()
list_raster_data_collections(params::Dict{String,<:Any})

Use this operation to get raster data collections.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The total number of items to return.

"NextToken": If the previous response was truncated, you receive this token. Use it in your next request to receive the next set of results.

Main.Sagemaker_Geospatial.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists the tags attached to the resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource you want to tag.

Main.Sagemaker_Geospatial.list_vector_enrichment_jobs — Method

list_vector_enrichment_jobs()
list_vector_enrichment_jobs(params::Dict{String,<:Any})

Retrieves a list of vector enrichment jobs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of items to return.

"NextToken": If the previous response was truncated, you receive this token. Use it in your next request to receive the next set of results.

"SortBy": The parameter by which to sort the results.

"SortOrder": An optional value that specifies whether you want the results sorted in Ascending or Descending order.

"StatusEquals": A filter that retrieves only jobs with a specific status.

Main.Sagemaker_Geospatial.search_raster_data_collection — Method

search_raster_data_collection(arn, raster_data_collection_query)
search_raster_data_collection(arn, raster_data_collection_query, params::Dict{String,<:Any})

Allows you run image query on a specific raster data collection to get a list of the satellite imagery matching the selected filters.

Arguments

arn: The Amazon Resource Name (ARN) of the raster data collection.

raster_data_collection_query:

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"NextToken": If the previous response was truncated, you receive this token. Use it in your next request to receive the next set of results.

Main.Sagemaker_Geospatial.start_earth_observation_job — Method

start_earth_observation_job(input_config, job_config, name)
start_earth_observation_job(input_config, job_config, name, params::Dict{String,<:Any})

Use this operation to create an Earth observation job.

Arguments

input_config: Input configuration information for the Earth Observation job.

job_config: An object containing information about the job configuration.

name: The name of the Earth Observation job.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": A unique token that guarantees that the call to this API is idempotent.

"ExecutionRoleArn": The Amazon Resource Name (ARN) of the IAM role that you specified for the job.

"KmsKeyId": The Amazon Key Management Service (KMS) key ID for server-side encryption.

"Tags": Each tag consists of a key and a value.

Main.Sagemaker_Geospatial.start_vector_enrichment_job — Method

start_vector_enrichment_job(execution_role_arn, input_config, job_config, name)
start_vector_enrichment_job(execution_role_arn, input_config, job_config, name, params::Dict{String,<:Any})

Creates a Vector Enrichment job for the supplied job type. Currently, there are two supported job types: reverse geocoding and map matching.

Arguments

execution_role_arn: The Amazon Resource Name (ARN) of the IAM role that you specified for the job.

input_config: Input configuration information for the Vector Enrichment job.

job_config: An object containing information about the job configuration.

name: The name of the Vector Enrichment job.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": A unique token that guarantees that the call to this API is idempotent.

"KmsKeyId": The Amazon Key Management Service (KMS) key ID for server-side encryption.

"Tags": Each tag consists of a key and a value.

Main.Sagemaker_Geospatial.stop_earth_observation_job — Method

stop_earth_observation_job(arn)
stop_earth_observation_job(arn, params::Dict{String,<:Any})

Use this operation to stop an existing earth observation job.

Arguments

arn: The Amazon Resource Name (ARN) of the Earth Observation job being stopped.

Main.Sagemaker_Geospatial.stop_vector_enrichment_job — Method

stop_vector_enrichment_job(arn)
stop_vector_enrichment_job(arn, params::Dict{String,<:Any})

Stops the Vector Enrichment job for a given job ARN.

Arguments

arn: The Amazon Resource Name (ARN) of the Vector Enrichment job.

Main.Sagemaker_Geospatial.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

The resource you want to tag.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource you want to tag.

tags: Each tag consists of a key and a value.

Main.Sagemaker_Geospatial.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

The resource you want to untag.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource you want to untag.

tag_keys: Keys of the tags you want to remove.