Schemas

This page documents function available when using the Schemas module, created with @service Schemas.

Index

Main.Schemas.create_discoverer

Main.Schemas.create_registry

Main.Schemas.create_schema

Main.Schemas.delete_discoverer

Main.Schemas.delete_registry

Main.Schemas.delete_resource_policy

Main.Schemas.delete_schema

Main.Schemas.delete_schema_version

Main.Schemas.describe_code_binding

Main.Schemas.describe_discoverer

Main.Schemas.describe_registry

Main.Schemas.describe_schema

Main.Schemas.export_schema

Main.Schemas.get_code_binding_source

Main.Schemas.get_discovered_schema

Main.Schemas.get_resource_policy

Main.Schemas.list_discoverers

Main.Schemas.list_registries

Main.Schemas.list_schema_versions

Main.Schemas.list_schemas

Main.Schemas.list_tags_for_resource

Main.Schemas.put_code_binding

Main.Schemas.put_resource_policy

Main.Schemas.search_schemas

Main.Schemas.start_discoverer

Main.Schemas.stop_discoverer

Main.Schemas.tag_resource

Main.Schemas.untag_resource

Main.Schemas.update_discoverer

Main.Schemas.update_registry

Main.Schemas.update_schema

Documentation

Main.Schemas.create_discoverer — Method

create_discoverer(source_arn)
create_discoverer(source_arn, params::Dict{String,<:Any})

Creates a discoverer.

Arguments

source_arn: The ARN of the event bus.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"CrossAccount": Support discovery of schemas in events sent to the bus from another account. (default: true).

"Description": A description for the discoverer.

"tags": Tags associated with the resource.

Main.Schemas.create_registry — Method

create_registry(registry_name)
create_registry(registry_name, params::Dict{String,<:Any})

Creates a registry.

Arguments

registry_name: The name of the registry.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A description of the registry to be created.

"tags": Tags to associate with the registry.

Main.Schemas.create_schema — Method

create_schema(content, type, registry_name, schema_name)
create_schema(content, type, registry_name, schema_name, params::Dict{String,<:Any})

Creates a schema definition. Inactive schemas will be deleted after two years.

Arguments

content: The source of the schema definition.

type: The type of schema.

registry_name: The name of the registry.

schema_name: The name of the schema.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A description of the schema.

"tags": Tags associated with the schema.

Main.Schemas.delete_discoverer — Method

delete_discoverer(discoverer_id)
delete_discoverer(discoverer_id, params::Dict{String,<:Any})

Deletes a discoverer.

Arguments

discoverer_id: The ID of the discoverer.

Main.Schemas.delete_registry — Method

delete_registry(registry_name)
delete_registry(registry_name, params::Dict{String,<:Any})

Deletes a Registry.

Arguments

registry_name: The name of the registry.

Main.Schemas.delete_resource_policy — Method

delete_resource_policy()
delete_resource_policy(params::Dict{String,<:Any})

Delete the resource-based policy attached to the specified registry.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"registryName": The name of the registry.

Main.Schemas.delete_schema — Method

delete_schema(registry_name, schema_name)
delete_schema(registry_name, schema_name, params::Dict{String,<:Any})

Delete a schema definition.

Arguments

registry_name: The name of the registry.

schema_name: The name of the schema.

Main.Schemas.delete_schema_version — Method

delete_schema_version(registry_name, schema_name, schema_version)
delete_schema_version(registry_name, schema_name, schema_version, params::Dict{String,<:Any})

Delete the schema version definition

Arguments

registry_name: The name of the registry.

schema_name: The name of the schema.

schema_version: The version number of the schema

Main.Schemas.describe_code_binding — Method

describe_code_binding(language, registry_name, schema_name)
describe_code_binding(language, registry_name, schema_name, params::Dict{String,<:Any})

Describe the code binding URI.

Arguments

language: The language of the code binding.

registry_name: The name of the registry.

schema_name: The name of the schema.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"schemaVersion": Specifying this limits the results to only this schema version.

Main.Schemas.describe_discoverer — Method

describe_discoverer(discoverer_id)
describe_discoverer(discoverer_id, params::Dict{String,<:Any})

Describes the discoverer.

Arguments

discoverer_id: The ID of the discoverer.

Main.Schemas.describe_registry — Method

describe_registry(registry_name)
describe_registry(registry_name, params::Dict{String,<:Any})

Describes the registry.

Arguments

registry_name: The name of the registry.

Main.Schemas.describe_schema — Method

describe_schema(registry_name, schema_name)
describe_schema(registry_name, schema_name, params::Dict{String,<:Any})

Retrieve the schema definition.

Arguments

registry_name: The name of the registry.

schema_name: The name of the schema.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"schemaVersion": Specifying this limits the results to only this schema version.

Main.Schemas.export_schema — Method

export_schema(registry_name, schema_name, type)
export_schema(registry_name, schema_name, type, params::Dict{String,<:Any})

Arguments

registry_name: The name of the registry.

schema_name: The name of the schema.

type:

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"schemaVersion": Specifying this limits the results to only this schema version.

Main.Schemas.get_code_binding_source — Method

get_code_binding_source(language, registry_name, schema_name)
get_code_binding_source(language, registry_name, schema_name, params::Dict{String,<:Any})

Get the code binding source URI.

Arguments

language: The language of the code binding.

registry_name: The name of the registry.

schema_name: The name of the schema.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"schemaVersion": Specifying this limits the results to only this schema version.

Main.Schemas.get_discovered_schema — Method

get_discovered_schema(events, type)
get_discovered_schema(events, type, params::Dict{String,<:Any})

Get the discovered schema that was generated based on sampled events.

Arguments

events: An array of strings where each string is a JSON event. These are the events that were used to generate the schema. The array includes a single type of event and has a maximum size of 10 events.

type: The type of event.

Main.Schemas.get_resource_policy — Method

get_resource_policy()
get_resource_policy(params::Dict{String,<:Any})

Retrieves the resource-based policy attached to a given registry.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"registryName": The name of the registry.

Main.Schemas.list_discoverers — Method

list_discoverers()
list_discoverers(params::Dict{String,<:Any})

List the discoverers.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"discovererIdPrefix": Specifying this limits the results to only those discoverer IDs that start with the specified prefix.

"limit":

"nextToken": The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

"sourceArnPrefix": Specifying this limits the results to only those ARNs that start with the specified prefix.

Main.Schemas.list_registries — Method

list_registries()
list_registries(params::Dict{String,<:Any})

List the registries.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"limit":

"nextToken": The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

"registryNamePrefix": Specifying this limits the results to only those registry names that start with the specified prefix.

"scope": Can be set to Local or AWS to limit responses to your custom registries, or the ones provided by AWS.

Main.Schemas.list_schema_versions — Method

list_schema_versions(registry_name, schema_name)
list_schema_versions(registry_name, schema_name, params::Dict{String,<:Any})

Provides a list of the schema versions and related information.

Arguments

registry_name: The name of the registry.

schema_name: The name of the schema.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"limit":

"nextToken": The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

Main.Schemas.list_schemas — Method

list_schemas(registry_name)
list_schemas(registry_name, params::Dict{String,<:Any})

List the schemas.

Arguments

registry_name: The name of the registry.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"limit":

"nextToken": The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

"schemaNamePrefix": Specifying this limits the results to only those schema names that start with the specified prefix.

Main.Schemas.list_tags_for_resource — Method

list_tags_for_resource(resource-arn)
list_tags_for_resource(resource-arn, params::Dict{String,<:Any})

Get tags for resource.

Arguments

resource-arn: The ARN of the resource.

Main.Schemas.put_code_binding — Method

put_code_binding(language, registry_name, schema_name)
put_code_binding(language, registry_name, schema_name, params::Dict{String,<:Any})

Put code binding URI

Arguments

language: The language of the code binding.

registry_name: The name of the registry.

schema_name: The name of the schema.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"schemaVersion": Specifying this limits the results to only this schema version.

Main.Schemas.put_resource_policy — Method

put_resource_policy(policy)
put_resource_policy(policy, params::Dict{String,<:Any})

The name of the policy.

Arguments

policy: The resource-based policy.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"RevisionId": The revision ID of the policy.

"registryName": The name of the registry.

Main.Schemas.search_schemas — Method

search_schemas(keywords, registry_name)
search_schemas(keywords, registry_name, params::Dict{String,<:Any})

Search the schemas

Arguments

keywords: Specifying this limits the results to only schemas that include the provided keywords.

registry_name: The name of the registry.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"limit":

"nextToken": The token that specifies the next page of results to return. To request the first page, leave NextToken empty. The token will expire in 24 hours, and cannot be shared with other accounts.

Main.Schemas.start_discoverer — Method

start_discoverer(discoverer_id)
start_discoverer(discoverer_id, params::Dict{String,<:Any})

Starts the discoverer

Arguments

discoverer_id: The ID of the discoverer.

Main.Schemas.stop_discoverer — Method

stop_discoverer(discoverer_id)
stop_discoverer(discoverer_id, params::Dict{String,<:Any})

Stops the discoverer

Arguments

discoverer_id: The ID of the discoverer.

Main.Schemas.tag_resource — Method

tag_resource(resource-arn, tags)
tag_resource(resource-arn, tags, params::Dict{String,<:Any})

Add tags to a resource.

Arguments

resource-arn: The ARN of the resource.

tags: Tags associated with the resource.

Main.Schemas.untag_resource — Method

untag_resource(resource-arn, tag_keys)
untag_resource(resource-arn, tag_keys, params::Dict{String,<:Any})

Removes tags from a resource.

Arguments

resource-arn: The ARN of the resource.

tag_keys: Keys of key-value pairs.

Main.Schemas.update_discoverer — Method

update_discoverer(discoverer_id)
update_discoverer(discoverer_id, params::Dict{String,<:Any})

Updates the discoverer

Arguments

discoverer_id: The ID of the discoverer.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"CrossAccount": Support discovery of schemas in events sent to the bus from another account. (default: true)

"Description": The description of the discoverer to update.

Main.Schemas.update_registry — Method

update_registry(registry_name)
update_registry(registry_name, params::Dict{String,<:Any})

Updates a registry.

Arguments

registry_name: The name of the registry.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": The description of the registry to update.

Main.Schemas.update_schema — Method

update_schema(registry_name, schema_name)
update_schema(registry_name, schema_name, params::Dict{String,<:Any})

Updates the schema definition Inactive schemas will be deleted after two years.

Arguments

registry_name: The name of the registry.

schema_name: The name of the schema.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientTokenId": The ID of the client token.

"Content": The source of the schema definition.

"Description": The description of the schema.

"Type": The schema type for the events schema.