Security Ir

This page documents function available when using the Security_Ir module, created with @service Security_Ir.

Index

Main.Security_Ir.batch_get_member_account_details

Main.Security_Ir.cancel_membership

Main.Security_Ir.close_case

Main.Security_Ir.create_case

Main.Security_Ir.create_case_comment

Main.Security_Ir.create_membership

Main.Security_Ir.get_case

Main.Security_Ir.get_case_attachment_download_url

Main.Security_Ir.get_case_attachment_upload_url

Main.Security_Ir.get_membership

Main.Security_Ir.list_case_edits

Main.Security_Ir.list_cases

Main.Security_Ir.list_comments

Main.Security_Ir.list_investigations

Main.Security_Ir.list_memberships

Main.Security_Ir.list_tags_for_resource

Main.Security_Ir.send_feedback

Main.Security_Ir.tag_resource

Main.Security_Ir.untag_resource

Main.Security_Ir.update_case

Main.Security_Ir.update_case_comment

Main.Security_Ir.update_case_status

Main.Security_Ir.update_membership

Main.Security_Ir.update_resolver_type

Documentation

Main.Security_Ir.batch_get_member_account_details — Function

batch_get_member_account_details(account_ids, membership_id)
batch_get_member_account_details(account_ids, membership_id, params::Dict{String,<:Any})

Provides information on whether the supplied account IDs are associated with a membership.

Note

AWS account ID's may appear less than 12 characters and need to be zero-prepended. An example would be 123123123 which is nine digits, and with zero-prepend would be 000123123123. Not zero-prepending to 12 digits could result in errors.

Arguments

account_ids: Optional element to query the membership relationship status to a provided list of account IDs.

Note

AWS account ID's may appear less than 12 characters and need to be zero-prepended. An example would be 123123123 which is nine digits, and with zero-prepend would be 000123123123. Not zero-prepending to 12 digits could result in errors.

membership_id: Required element used in combination with BatchGetMemberAccountDetails to identify the membership ID to query.

Main.Security_Ir.cancel_membership — Function

cancel_membership(membership_id)
cancel_membership(membership_id, params::Dict{String,<:Any})

Cancels an existing membership.

Arguments

membership_id: Required element used in combination with CancelMembershipRequest to identify the membership ID to cancel.

Main.Security_Ir.close_case — Function

close_case(case_id)
close_case(case_id, params::Dict{String,<:Any})

Closes an existing case.

Arguments

case_id: Required element used in combination with CloseCase to identify the case ID to close.

Main.Security_Ir.create_case — Function

create_case(description, engagement_type, impacted_accounts, reported_incident_start_date, resolver_type, title, watchers)
create_case(description, engagement_type, impacted_accounts, reported_incident_start_date, resolver_type, title, watchers, params::Dict{String,<:Any})

Creates a new case.

Arguments

description: Required element used in combination with CreateCase

to provide a description for the new case.

engagement_type: Required element used in combination with CreateCase to provide an engagement type for the new cases. Available engagement types include Security Incident | Investigation

impacted_accounts: Required element used in combination with CreateCase to provide a list of impacted accounts.

Note

AWS account ID's may appear less than 12 characters and need to be zero-prepended. An example would be 123123123 which is nine digits, and with zero-prepend would be 000123123123. Not zero-prepending to 12 digits could result in errors.

reported_incident_start_date: Required element used in combination with CreateCase to provide an initial start date for the unauthorized activity.

resolver_type: Required element used in combination with CreateCase to identify the resolver type.

title: Required element used in combination with CreateCase to provide a title for the new case.

watchers: Required element used in combination with CreateCase to provide a list of entities to receive notifications for case updates.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": !!! note The clientToken field is an idempotency key used to ensure that repeated attempts for a single action will be ignored by the server during retries. A caller supplied unique ID (typically a UUID) should be provided.

"impactedAwsRegions": An optional element used in combination with CreateCase to provide a list of impacted regions.

"impactedServices": An optional element used in combination with CreateCase to provide a list of services impacted.

"tags": An optional element used in combination with CreateCase to add customer specified tags to a case.

"threatActorIpAddresses": An optional element used in combination with CreateCase to provide a list of suspicious internet protocol addresses associated with unauthorized activity.

Main.Security_Ir.create_case_comment — Function

create_case_comment(body, case_id)
create_case_comment(body, case_id, params::Dict{String,<:Any})

Adds a comment to an existing case.

Arguments

body: Required element used in combination with CreateCaseComment to add content for the new comment.

case_id: Required element used in combination with CreateCaseComment to specify a case ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": !!! note The clientToken field is an idempotency key used to ensure that repeated attempts for a single action will be ignored by the server during retries. A caller supplied unique ID (typically a UUID) should be provided.

Main.Security_Ir.create_membership — Function

create_membership(incident_response_team, membership_name)
create_membership(incident_response_team, membership_name, params::Dict{String,<:Any})

Creates a new membership.

Arguments

incident_response_team: Required element used in combination with CreateMembership to add customer incident response team members and trusted partners to the membership.

membership_name: Required element used in combination with CreateMembership to create a name for the membership.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": !!! note The clientToken field is an idempotency key used to ensure that repeated attempts for a single action will be ignored by the server during retries. A caller supplied unique ID (typically a UUID) should be provided.

"coverEntireOrganization": The coverEntireOrganization parameter is a boolean flag that determines whether the membership should be applied to the entire Amazon Web Services Organization. When set to true, the membership will be created for all accounts within the organization. When set to false, the membership will only be created for specified accounts.

This parameter is optional. If not specified, the default value is false.

If set to true: The membership will automatically include all existing and future accounts in the Amazon Web Services Organization.

If set to false: The membership will only apply to explicitly specified accounts.

"optInFeatures": Optional element to enable the monitoring and investigation opt-in features for the service.

"tags": Optional element for customer configured tags.

Main.Security_Ir.get_case — Function

get_case(case_id)
get_case(case_id, params::Dict{String,<:Any})

Returns the attributes of a case.

Arguments

case_id: Required element for GetCase to identify the requested case ID.

Main.Security_Ir.get_case_attachment_download_url — Function

get_case_attachment_download_url(attachment_id, case_id)
get_case_attachment_download_url(attachment_id, case_id, params::Dict{String,<:Any})

Returns a Pre-Signed URL for uploading attachments into a case.

Arguments

attachment_id: Required element for GetCaseAttachmentDownloadUrl to identify the attachment ID for downloading an attachment.

case_id: Required element for GetCaseAttachmentDownloadUrl to identify the case ID for downloading an attachment from.

Main.Security_Ir.get_case_attachment_upload_url — Function

get_case_attachment_upload_url(case_id, content_length, file_name)
get_case_attachment_upload_url(case_id, content_length, file_name, params::Dict{String,<:Any})

Uploads an attachment to a case.

Arguments

case_id: Required element for GetCaseAttachmentUploadUrl to identify the case ID for uploading an attachment.

content_length: Required element for GetCaseAttachmentUploadUrl to identify the size of the file attachment.

file_name: Required element for GetCaseAttachmentUploadUrl to identify the file name of the attachment to upload.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": !!! note The clientToken field is an idempotency key used to ensure that repeated attempts for a single action will be ignored by the server during retries. A caller supplied unique ID (typically a UUID) should be provided.

Main.Security_Ir.get_membership — Function

get_membership(membership_id)
get_membership(membership_id, params::Dict{String,<:Any})

Returns the attributes of a membership.

Arguments

membership_id: Required element for GetMembership to identify the membership ID to query.

Main.Security_Ir.list_case_edits — Function

list_case_edits(case_id)
list_case_edits(case_id, params::Dict{String,<:Any})

Views the case history for edits made to a designated case.

Arguments

case_id: Required element used with ListCaseEdits to identify the case to query.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Optional element to identify how many results to obtain. There is a maximum value of 25.

"nextToken": An optional string that, if supplied, must be copied from the output of a previous call to ListCaseEdits. When provided in this manner, the API fetches the next page of results.

Main.Security_Ir.list_cases — Function

list_cases()
list_cases(params::Dict{String,<:Any})

Lists all cases the requester has access to.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Optional element for ListCases to limit the number of responses.

"nextToken": An optional string that, if supplied, must be copied from the output of a previous call to ListCases. When provided in this manner, the API fetches the next page of results.

Main.Security_Ir.list_comments — Function

list_comments(case_id)
list_comments(case_id, params::Dict{String,<:Any})

Returns comments for a designated case.

Arguments

case_id: Required element for ListComments to designate the case to query.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Optional element for ListComments to limit the number of responses.

"nextToken": An optional string that, if supplied, must be copied from the output of a previous call to ListComments. When provided in this manner, the API fetches the next page of results.

Main.Security_Ir.list_investigations — Function

list_investigations(case_id)
list_investigations(case_id, params::Dict{String,<:Any})

Investigation performed by an agent for a security incident...

Arguments

case_id: Investigation performed by an agent for a security incident per caseID

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Investigation performed by an agent for a security incident request, returning max results

"nextToken": Investigation performed by an agent for a security incident request

Main.Security_Ir.list_memberships — Function

list_memberships()
list_memberships(params::Dict{String,<:Any})

Returns the memberships that the calling principal can access.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Request element for ListMemberships to limit the number of responses.

"nextToken": An optional string that, if supplied, must be copied from the output of a previous call to ListMemberships. When provided in this manner, the API fetches the next page of results.

Main.Security_Ir.list_tags_for_resource — Function

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Returns currently configured tags on a resource.

Arguments

resource_arn: Required element for ListTagsForResource to provide the ARN to identify a specific resource.

Main.Security_Ir.send_feedback — Function

send_feedback(case_id, result_id, usefulness)
send_feedback(case_id, result_id, usefulness, params::Dict{String,<:Any})

Send feedback based on response investigation action

Arguments

case_id: Send feedback based on request caseID

result_id: Send feedback based on request result ID

usefulness: Required enum value indicating user assessment of result q.....

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"comment": Send feedback based on request comments

Main.Security_Ir.tag_resource — Function

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Adds a tag(s) to a designated resource.

Arguments

resource_arn: Required element for TagResource to identify the ARN for the resource to add a tag to.

tags: Required element for ListTagsForResource to provide the content for a tag.

Main.Security_Ir.untag_resource — Function

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes a tag(s) from a designate resource.

Arguments

resource_arn: Required element for UnTagResource to identify the ARN for the resource to remove a tag from.

tag_keys: Required element for UnTagResource to identify tag to remove.

Main.Security_Ir.update_case — Function

update_case(case_id)
update_case(case_id, params::Dict{String,<:Any})

Updates an existing case.

Arguments

case_id: Required element for UpdateCase to identify the case ID for updates.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"actualIncidentStartDate": Optional element for UpdateCase to provide content for the incident start date field.

"caseMetadata": Update the case request with case metadata

"description": Optional element for UpdateCase to provide content for the description field.

"engagementType": Optional element for UpdateCase to provide content for the engagement type field. Available engagement types include Security Incident | Investigation.

"impactedAccountsToAdd": Optional element for UpdateCase to provide content to add accounts impacted.

Note

AWS account ID's may appear less than 12 characters and need to be zero-prepended. An example would be 123123123 which is nine digits, and with zero-prepend would be 000123123123. Not zero-prepending to 12 digits could result in errors.

"impactedAccountsToDelete": Optional element for UpdateCase to provide content to add accounts impacted.

Note

AWS account ID's may appear less than 12 characters and need to be zero-prepended. An example would be 123123123 which is nine digits, and with zero-prepend would be 000123123123. Not zero-prepending to 12 digits could result in errors.

"impactedAwsRegionsToAdd": Optional element for UpdateCase to provide content to add regions impacted.

"impactedAwsRegionsToDelete": Optional element for UpdateCase to provide content to remove regions impacted.

"impactedServicesToAdd": Optional element for UpdateCase to provide content to add services impacted.

"impactedServicesToDelete": Optional element for UpdateCase to provide content to remove services impacted.

"reportedIncidentStartDate": Optional element for UpdateCase to provide content for the customer reported incident start date field.

"threatActorIpAddressesToAdd": Optional element for UpdateCase to provide content to add additional suspicious IP addresses related to a case.

"threatActorIpAddressesToDelete": Optional element for UpdateCase to provide content to remove suspicious IP addresses from a case.

"title": Optional element for UpdateCase to provide content for the title field.

"watchersToAdd": Optional element for UpdateCase to provide content to add additional watchers to a case.

"watchersToDelete": Optional element for UpdateCase to provide content to remove existing watchers from a case.

Main.Security_Ir.update_case_comment — Function

update_case_comment(body, case_id, comment_id)
update_case_comment(body, case_id, comment_id, params::Dict{String,<:Any})

Updates an existing case comment.

Arguments

body: Required element for UpdateCaseComment to identify the content for the comment to be updated.

case_id: Required element for UpdateCaseComment to identify the case ID containing the comment to be updated.

comment_id: Required element for UpdateCaseComment to identify the case ID to be updated.

Main.Security_Ir.update_case_status — Function

update_case_status(case_id, case_status)
update_case_status(case_id, case_status, params::Dict{String,<:Any})

Updates the state transitions for a designated cases.

Self-managed: the following states are available for self-managed cases.

Submitted → Detection and Analysis

Detection and Analysis → Containment, Eradication, and Recovery

Detection and Analysis → Post-incident Activities

Containment, Eradication, and Recovery → Detection and Analysis

Containment, Eradication, and Recovery → Post-incident Activities

Post-incident Activities → Containment, Eradication, and Recovery

Post-incident Activities → Detection and Analysis

Any → Closed

AWS supported: You must use the CloseCase API to close.

Arguments

case_id: Required element for UpdateCaseStatus to identify the case to update.

case_status: Required element for UpdateCaseStatus to identify the status for a case. Options include Submitted | Detection and Analysis | Containment, Eradication and Recovery | Post-incident Activities.

Main.Security_Ir.update_membership — Function

update_membership(membership_id)
update_membership(membership_id, params::Dict{String,<:Any})

Updates membership configuration.

Arguments

membership_id: Required element for UpdateMembership to identify the membership to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"incidentResponseTeam": Optional element for UpdateMembership to update the membership name.

"membershipAccountsConfigurationsUpdate": The membershipAccountsConfigurationsUpdate field in the UpdateMembershipRequest structure allows you to update the configuration settings for accounts within a membership.

This field is optional and contains a structure of type MembershipAccountsConfigurationsUpdate that specifies the updated account configurations for the membership.

"membershipName": Optional element for UpdateMembership to update the membership name.

"optInFeatures": Optional element for UpdateMembership to enable or disable opt-in features for the service.

"undoMembershipCancellation": The undoMembershipCancellation parameter is a boolean flag that indicates whether to reverse a previously requested membership cancellation. When set to true, this will revoke the cancellation request and maintain the membership status.

This parameter is optional and can be used in scenarios where you need to restore a membership that was marked for cancellation but hasn't been fully terminated yet.

If set to true, the cancellation request will be revoked

If set to false the service will throw a ValidationException.

Main.Security_Ir.update_resolver_type — Function

update_resolver_type(case_id, resolver_type)
update_resolver_type(case_id, resolver_type, params::Dict{String,<:Any})

Updates the resolver type for a case.

Important

This is a one-way action and cannot be reversed.

Arguments

case_id: Required element for UpdateResolverType to identify the case to update.

resolver_type: Required element for UpdateResolverType to identify the new resolver.