Securityhub

This page documents function available when using the Securityhub module, created with @service Securityhub.

Index

Main.Securityhub.accept_administrator_invitation

Main.Securityhub.accept_invitation

Main.Securityhub.batch_disable_standards

Main.Securityhub.batch_enable_standards

Main.Securityhub.batch_import_findings

Main.Securityhub.batch_update_findings

Main.Securityhub.create_action_target

Main.Securityhub.create_insight

Main.Securityhub.create_members

Main.Securityhub.decline_invitations

Main.Securityhub.delete_action_target

Main.Securityhub.delete_insight

Main.Securityhub.delete_invitations

Main.Securityhub.delete_members

Main.Securityhub.describe_action_targets

Main.Securityhub.describe_hub

Main.Securityhub.describe_organization_configuration

Main.Securityhub.describe_products

Main.Securityhub.describe_standards

Main.Securityhub.describe_standards_controls

Main.Securityhub.disable_import_findings_for_product

Main.Securityhub.disable_organization_admin_account

Main.Securityhub.disable_security_hub

Main.Securityhub.disassociate_from_administrator_account

Main.Securityhub.disassociate_from_master_account

Main.Securityhub.disassociate_members

Main.Securityhub.enable_import_findings_for_product

Main.Securityhub.enable_organization_admin_account

Main.Securityhub.enable_security_hub

Main.Securityhub.get_administrator_account

Main.Securityhub.get_enabled_standards

Main.Securityhub.get_findings

Main.Securityhub.get_insight_results

Main.Securityhub.get_insights

Main.Securityhub.get_invitations_count

Main.Securityhub.get_master_account

Main.Securityhub.get_members

Main.Securityhub.invite_members

Main.Securityhub.list_enabled_products_for_import

Main.Securityhub.list_invitations

Main.Securityhub.list_members

Main.Securityhub.list_organization_admin_accounts

Main.Securityhub.list_tags_for_resource

Main.Securityhub.tag_resource

Main.Securityhub.untag_resource

Main.Securityhub.update_action_target

Main.Securityhub.update_findings

Main.Securityhub.update_insight

Main.Securityhub.update_organization_configuration

Main.Securityhub.update_security_hub_configuration

Main.Securityhub.update_standards_control

Documentation

Main.Securityhub.accept_administrator_invitation — Method

accept_administrator_invitation(administrator_id, invitation_id)
accept_administrator_invitation(administrator_id, invitation_id, params::Dict{String,<:Any})

Accepts the invitation to be a member account and be monitored by the Security Hub administrator account that the invitation was sent from. This operation is only used by member accounts that are not added through Organizations. When the member account accepts the invitation, permission is granted to the administrator account to view findings generated in the member account.

Arguments

administrator_id: The account ID of the Security Hub administrator account that sent the invitation.

invitation_id: The identifier of the invitation sent from the Security Hub administrator account.

Main.Securityhub.accept_invitation — Method

accept_invitation(invitation_id, master_id)
accept_invitation(invitation_id, master_id, params::Dict{String,<:Any})

This method is deprecated. Instead, use AcceptAdministratorInvitation. Accepts the invitation to be a member account and be monitored by the Security Hub administrator account that the invitation was sent from. This operation is only used by member accounts that are not added through Organizations. When the member account accepts the invitation, permission is granted to the administrator account to view findings generated in the member account.

Arguments

invitation_id: The identifier of the invitation sent from the Security Hub administrator account.

master_id: The account ID of the Security Hub administrator account that sent the invitation.

Main.Securityhub.batch_disable_standards — Method

batch_disable_standards(standards_subscription_arns)
batch_disable_standards(standards_subscription_arns, params::Dict{String,<:Any})

Disables the standards specified by the provided StandardsSubscriptionArns. For more information, see Security Standards section of the AWS Security Hub User Guide.

Arguments

standards_subscription_arns: The ARNs of the standards subscriptions to disable.

Main.Securityhub.batch_enable_standards — Method

batch_enable_standards(standards_subscription_requests)
batch_enable_standards(standards_subscription_requests, params::Dict{String,<:Any})

Enables the standards specified by the provided StandardsArn. To obtain the ARN for a standard, use the DescribeStandards operation. For more information, see the Security Standards section of the AWS Security Hub User Guide.

Arguments

standards_subscription_requests: The list of standards checks to enable.

Main.Securityhub.batch_import_findings — Method

batch_import_findings(findings)
batch_import_findings(findings, params::Dict{String,<:Any})

Imports security findings generated from an integrated product into Security Hub. This action is requested by the integrated product to import its findings into Security Hub. The maximum allowed size for a finding is 240 Kb. An error is returned for any finding larger than 240 Kb. After a finding is created, BatchImportFindings cannot be used to update the following finding fields and objects, which Security Hub customers use to manage their investigation workflow. Note UserDefinedFields VerificationState Workflow Finding providers also should not use BatchImportFindings to update the following attributes. Confidence Criticality RelatedFindings Severity Types Instead, finding providers use FindingProviderFields to provide values for these attributes.

Arguments

findings: A list of findings to import. To successfully import a finding, it must follow the AWS Security Finding Format. Maximum of 100 findings per request.

Main.Securityhub.batch_update_findings — Method

batch_update_findings(finding_identifiers)
batch_update_findings(finding_identifiers, params::Dict{String,<:Any})

Used by Security Hub customers to update information about their investigation into a finding. Requested by administrator accounts or member accounts. Administrator accounts can update findings for their account and their member accounts. Member accounts can update findings for their account. Updates from BatchUpdateFindings do not affect the value of UpdatedAt for a finding. Administrator and member accounts can use BatchUpdateFindings to update the following finding fields and objects. Confidence Criticality Note RelatedFindings Severity Types UserDefinedFields VerificationState Workflow You can configure IAM policies to restrict access to fields and field values. For example, you might not want member accounts to be able to suppress findings or change the finding severity. See Configuring access to BatchUpdateFindings in the AWS Security Hub User Guide.

Arguments

finding_identifiers: The list of findings to update. BatchUpdateFindings can be used to update up to 100 findings at a time. For each finding, the list provides the finding identifier and the ARN of the finding provider.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Confidence": The updated value for the finding confidence. Confidence is defined as the likelihood that a finding accurately identifies the behavior or issue that it was intended to identify. Confidence is scored on a 0-100 basis using a ratio scale, where 0 means zero percent confidence and 100 means 100 percent confidence.

"Criticality": The updated value for the level of importance assigned to the resources associated with the findings. A score of 0 means that the underlying resources have no criticality, and a score of 100 is reserved for the most critical resources.

"Note":

"RelatedFindings": A list of findings that are related to the updated findings.

"Severity": Used to update the finding severity.

"Types": One or more finding types in the format of namespace/category/classifier that classify a finding. Valid namespace values are as follows. Software and Configuration Checks TTPs Effects Unusual Behaviors Sensitive Data Identifications

"UserDefinedFields": A list of name/value string pairs associated with the finding. These are custom, user-defined fields added to a finding.

"VerificationState": Indicates the veracity of a finding. The available values for VerificationState are as follows. UNKNOWN – The default disposition of a security finding TRUEPOSITIVE – The security finding is confirmed FALSEPOSITIVE – The security finding was determined to be a false alarm BENIGNPOSITIVE – A special case of TRUEPOSITIVE where the finding doesn't pose any threat, is expected, or both

"Workflow": Used to update the workflow status of a finding. The workflow status indicates the progress of the investigation into the finding.

Main.Securityhub.create_action_target — Method

create_action_target(description, id, name)
create_action_target(description, id, name, params::Dict{String,<:Any})

Creates a custom action target in Security Hub. You can use custom actions on findings and insights in Security Hub to trigger target actions in Amazon CloudWatch Events.

Arguments

description: The description for the custom action target.

id: The ID for the custom action target.

name: The name of the custom action target.

Main.Securityhub.create_insight — Method

create_insight(filters, group_by_attribute, name)
create_insight(filters, group_by_attribute, name, params::Dict{String,<:Any})

Creates a custom insight in Security Hub. An insight is a consolidation of findings that relate to a security issue that requires attention or remediation. To group the related findings in the insight, use the GroupByAttribute.

Arguments

filters: One or more attributes used to filter the findings included in the insight. The insight only includes findings that match the criteria defined in the filters.

group_by_attribute: The attribute used to group the findings for the insight. The grouping attribute identifies the type of item that the insight applies to. For example, if an insight is grouped by resource identifier, then the insight produces a list of resource identifiers.

name: The name of the custom insight to create.

Main.Securityhub.create_members — Method

create_members(account_details)
create_members(account_details, params::Dict{String,<:Any})

Creates a member association in Security Hub between the specified accounts and the account used to make the request, which is the administrator account. If you are integrated with Organizations, then the administrator account is designated by the organization management account. CreateMembers is always used to add accounts that are not organization members. For accounts that are part of an organization, CreateMembers is only used in the following cases: Security Hub is not configured to automatically add new accounts in an organization. The account was disassociated or deleted in Security Hub. This action can only be used by an account that has Security Hub enabled. To enable Security Hub, you can use the EnableSecurityHub operation. For accounts that are not organization members, you create the account association and then send an invitation to the member account. To send the invitation, you use the InviteMembers operation. If the account owner accepts the invitation, the account becomes a member account in Security Hub. Accounts that are part of an organization do not receive an invitation. They automatically become a member account in Security Hub. A permissions policy is added that permits the administrator account to view the findings generated in the member account. When Security Hub is enabled in a member account, the member account findings are also visible to the administrator account. To remove the association between the administrator and member accounts, use the DisassociateFromMasterAccount or DisassociateMembers operation.

Arguments

account_details: The list of accounts to associate with the Security Hub administrator account. For each account, the list includes the account ID and optionally the email address.

Main.Securityhub.decline_invitations — Method

decline_invitations(account_ids)
decline_invitations(account_ids, params::Dict{String,<:Any})

Declines invitations to become a member account. This operation is only used by accounts that are not part of an organization. Organization accounts do not receive invitations.

Arguments

account_ids: The list of account IDs for the accounts from which to decline the invitations to Security Hub.

Main.Securityhub.delete_action_target — Method

delete_action_target(action_target_arn)
delete_action_target(action_target_arn, params::Dict{String,<:Any})

Deletes a custom action target from Security Hub. Deleting a custom action target does not affect any findings or insights that were already sent to Amazon CloudWatch Events using the custom action.

Arguments

action_target_arn: The ARN of the custom action target to delete.

Main.Securityhub.delete_insight — Method

delete_insight(insight_arn)
delete_insight(insight_arn, params::Dict{String,<:Any})

Deletes the insight specified by the InsightArn.

Arguments

insight_arn: The ARN of the insight to delete.

Main.Securityhub.delete_invitations — Method

delete_invitations(account_ids)
delete_invitations(account_ids, params::Dict{String,<:Any})

Deletes invitations received by the AWS account to become a member account. This operation is only used by accounts that are not part of an organization. Organization accounts do not receive invitations.

Arguments

account_ids: The list of the account IDs that sent the invitations to delete.

Main.Securityhub.delete_members — Method

delete_members(account_ids)
delete_members(account_ids, params::Dict{String,<:Any})

Deletes the specified member accounts from Security Hub. Can be used to delete member accounts that belong to an organization as well as member accounts that were invited manually.

Arguments

account_ids: The list of account IDs for the member accounts to delete.

Main.Securityhub.describe_action_targets — Method

describe_action_targets()
describe_action_targets(params::Dict{String,<:Any})

Returns a list of the custom action targets in Security Hub in your account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ActionTargetArns": A list of custom action target ARNs for the custom action targets to retrieve.

"MaxResults": The maximum number of results to return.

"NextToken": The token that is required for pagination. On your first call to the DescribeActionTargets operation, set the value of this parameter to NULL. For subsequent calls to the operation, to continue listing data, set the value of this parameter to the value returned from the previous response.

Main.Securityhub.describe_hub — Method

describe_hub()
describe_hub(params::Dict{String,<:Any})

Returns details about the Hub resource in your account, including the HubArn and the time when you enabled Security Hub.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"HubArn": The ARN of the Hub resource to retrieve.

Main.Securityhub.describe_organization_configuration — Method

describe_organization_configuration()
describe_organization_configuration(params::Dict{String,<:Any})

Returns information about the Organizations configuration for Security Hub. Can only be called from a Security Hub administrator account.

Main.Securityhub.describe_products — Method

describe_products()
describe_products(params::Dict{String,<:Any})

Returns information about product integrations in Security Hub. You can optionally provide an integration ARN. If you provide an integration ARN, then the results only include that integration. If you do not provide an integration ARN, then the results include all of the available product integrations.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return.

"NextToken": The token that is required for pagination. On your first call to the DescribeProducts operation, set the value of this parameter to NULL. For subsequent calls to the operation, to continue listing data, set the value of this parameter to the value returned from the previous response.

"ProductArn": The ARN of the integration to return.

Main.Securityhub.describe_standards — Method

describe_standards()
describe_standards(params::Dict{String,<:Any})

Returns a list of the available standards in Security Hub. For each standard, the results include the standard ARN, the name, and a description.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of standards to return.

"NextToken": The token that is required for pagination. On your first call to the DescribeStandards operation, set the value of this parameter to NULL. For subsequent calls to the operation, to continue listing data, set the value of this parameter to the value returned from the previous response.

Main.Securityhub.describe_standards_controls — Method

describe_standards_controls(standards_subscription_arn)
describe_standards_controls(standards_subscription_arn, params::Dict{String,<:Any})

Returns a list of security standards controls. For each control, the results include information about whether it is currently enabled, the severity, and a link to remediation information.

Arguments

standards_subscription_arn: The ARN of a resource that represents your subscription to a supported standard. To get the subscription ARNs of the standards you have enabled, use the GetEnabledStandards operation.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of security standard controls to return.

"NextToken": The token that is required for pagination. On your first call to the DescribeStandardsControls operation, set the value of this parameter to NULL. For subsequent calls to the operation, to continue listing data, set the value of this parameter to the value returned from the previous response.

Main.Securityhub.disable_import_findings_for_product — Method

disable_import_findings_for_product(product_subscription_arn)
disable_import_findings_for_product(product_subscription_arn, params::Dict{String,<:Any})

Disables the integration of the specified product with Security Hub. After the integration is disabled, findings from that product are no longer sent to Security Hub.

Arguments

product_subscription_arn: The ARN of the integrated product to disable the integration for.

Main.Securityhub.disable_organization_admin_account — Method

disable_organization_admin_account(admin_account_id)
disable_organization_admin_account(admin_account_id, params::Dict{String,<:Any})

Disables a Security Hub administrator account. Can only be called by the organization management account.

Arguments

admin_account_id: The AWS account identifier of the Security Hub administrator account.

Main.Securityhub.disable_security_hub — Method

disable_security_hub()
disable_security_hub(params::Dict{String,<:Any})

Disables Security Hub in your account only in the current Region. To disable Security Hub in all Regions, you must submit one request per Region where you have enabled Security Hub. When you disable Security Hub for an administrator account, it doesn't disable Security Hub for any associated member accounts. When you disable Security Hub, your existing findings and insights and any Security Hub configuration settings are deleted after 90 days and cannot be recovered. Any standards that were enabled are disabled, and your administrator and member account associations are removed. If you want to save your existing findings, you must export them before you disable Security Hub.

Main.Securityhub.disassociate_from_administrator_account — Method

disassociate_from_administrator_account()
disassociate_from_administrator_account(params::Dict{String,<:Any})

Disassociates the current Security Hub member account from the associated administrator account. This operation is only used by accounts that are not part of an organization. For organization accounts, only the administrator account can disassociate a member account.

Main.Securityhub.disassociate_from_master_account — Method

disassociate_from_master_account()
disassociate_from_master_account(params::Dict{String,<:Any})

This method is deprecated. Instead, use DisassociateFromAdministratorAccount. Disassociates the current Security Hub member account from the associated administrator account. This operation is only used by accounts that are not part of an organization. For organization accounts, only the administrator account can disassociate a member account.

Main.Securityhub.disassociate_members — Method

disassociate_members(account_ids)
disassociate_members(account_ids, params::Dict{String,<:Any})

Disassociates the specified member accounts from the associated administrator account. Can be used to disassociate both accounts that are managed using Organizations and accounts that were invited manually.

Arguments

account_ids: The account IDs of the member accounts to disassociate from the administrator account.

Main.Securityhub.enable_import_findings_for_product — Method

enable_import_findings_for_product(product_arn)
enable_import_findings_for_product(product_arn, params::Dict{String,<:Any})

Enables the integration of a partner product with Security Hub. Integrated products send findings to Security Hub. When you enable a product integration, a permissions policy that grants permission for the product to send findings to Security Hub is applied.

Arguments

product_arn: The ARN of the product to enable the integration for.

Main.Securityhub.enable_organization_admin_account — Method

enable_organization_admin_account(admin_account_id)
enable_organization_admin_account(admin_account_id, params::Dict{String,<:Any})

Designates the Security Hub administrator account for an organization. Can only be called by the organization management account.

Arguments

admin_account_id: The AWS account identifier of the account to designate as the Security Hub administrator account.

Main.Securityhub.enable_security_hub — Method

enable_security_hub()
enable_security_hub(params::Dict{String,<:Any})

Enables Security Hub for your account in the current Region or the Region you specify in the request. When you enable Security Hub, you grant to Security Hub the permissions necessary to gather findings from other services that are integrated with Security Hub. When you use the EnableSecurityHub operation to enable Security Hub, you also automatically enable the following standards. CIS AWS Foundations AWS Foundational Security Best Practices You do not enable the Payment Card Industry Data Security Standard (PCI DSS) standard. To not enable the automatically enabled standards, set EnableDefaultStandards to false. After you enable Security Hub, to enable a standard, use the BatchEnableStandards operation. To disable a standard, use the BatchDisableStandards operation. To learn more, see Setting Up AWS Security Hub in the AWS Security Hub User Guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"EnableDefaultStandards": Whether to enable the security standards that Security Hub has designated as automatically enabled. If you do not provide a value for EnableDefaultStandards, it is set to true. To not enable the automatically enabled standards, set EnableDefaultStandards to false.

"Tags": The tags to add to the hub resource when you enable Security Hub.

Main.Securityhub.get_administrator_account — Method

get_administrator_account()
get_administrator_account(params::Dict{String,<:Any})

Provides the details for the Security Hub administrator account for the current member account. Can be used by both member accounts that are managed using Organizations and accounts that were invited manually.

Main.Securityhub.get_enabled_standards — Method

get_enabled_standards()
get_enabled_standards(params::Dict{String,<:Any})

Returns a list of the standards that are currently enabled.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return in the response.

"NextToken": The token that is required for pagination. On your first call to the GetEnabledStandards operation, set the value of this parameter to NULL. For subsequent calls to the operation, to continue listing data, set the value of this parameter to the value returned from the previous response.

"StandardsSubscriptionArns": The list of the standards subscription ARNs for the standards to retrieve.

Main.Securityhub.get_findings — Method

get_findings()
get_findings(params::Dict{String,<:Any})

Returns a list of findings that match the specified criteria.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Filters": The finding attributes used to define a condition to filter the returned findings. You can filter by up to 10 finding attributes. For each attribute, you can provide up to 20 filter values. Note that in the available filter fields, WorkflowState is deprecated. To search for a finding based on its workflow status, use WorkflowStatus.

"MaxResults": The maximum number of findings to return.

"NextToken": The token that is required for pagination. On your first call to the GetFindings operation, set the value of this parameter to NULL. For subsequent calls to the operation, to continue listing data, set the value of this parameter to the value returned from the previous response.

"SortCriteria": The finding attributes used to sort the list of returned findings.

Main.Securityhub.get_insight_results — Method

get_insight_results(insight_arn)
get_insight_results(insight_arn, params::Dict{String,<:Any})

Lists the results of the Security Hub insight specified by the insight ARN.

Arguments

insight_arn: The ARN of the insight for which to return results.

Main.Securityhub.get_insights — Method

get_insights()
get_insights(params::Dict{String,<:Any})

Lists and describes insights for the specified insight ARNs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"InsightArns": The ARNs of the insights to describe. If you do not provide any insight ARNs, then GetInsights returns all of your custom insights. It does not return any managed insights.

"MaxResults": The maximum number of items to return in the response.

"NextToken": The token that is required for pagination. On your first call to the GetInsights operation, set the value of this parameter to NULL. For subsequent calls to the operation, to continue listing data, set the value of this parameter to the value returned from the previous response.

Main.Securityhub.get_invitations_count — Method

get_invitations_count()
get_invitations_count(params::Dict{String,<:Any})

Returns the count of all Security Hub membership invitations that were sent to the current member account, not including the currently accepted invitation.

Main.Securityhub.get_master_account — Method

get_master_account()
get_master_account(params::Dict{String,<:Any})

This method is deprecated. Instead, use GetAdministratorAccount. Provides the details for the Security Hub administrator account for the current member account. Can be used by both member accounts that are managed using Organizations and accounts that were invited manually.

Main.Securityhub.get_members — Method

get_members(account_ids)
get_members(account_ids, params::Dict{String,<:Any})

Returns the details for the Security Hub member accounts for the specified account IDs. An administrator account can be either the delegated Security Hub administrator account for an organization or an administrator account that enabled Security Hub manually. The results include both member accounts that are managed using Organizations and accounts that were invited manually.

Arguments

account_ids: The list of account IDs for the Security Hub member accounts to return the details for.

Main.Securityhub.invite_members — Method

invite_members(account_ids)
invite_members(account_ids, params::Dict{String,<:Any})

Invites other AWS accounts to become member accounts for the Security Hub administrator account that the invitation is sent from. This operation is only used to invite accounts that do not belong to an organization. Organization accounts do not receive invitations. Before you can use this action to invite a member, you must first use the CreateMembers action to create the member account in Security Hub. When the account owner enables Security Hub and accepts the invitation to become a member account, the administrator account can view the findings generated from the member account.

Arguments

account_ids: The list of account IDs of the AWS accounts to invite to Security Hub as members.

Main.Securityhub.list_enabled_products_for_import — Method

list_enabled_products_for_import()
list_enabled_products_for_import(params::Dict{String,<:Any})

Lists all findings-generating solutions (products) that you are subscribed to receive findings from in Security Hub.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of items to return in the response.

"NextToken": The token that is required for pagination. On your first call to the ListEnabledProductsForImport operation, set the value of this parameter to NULL. For subsequent calls to the operation, to continue listing data, set the value of this parameter to the value returned from the previous response.

Main.Securityhub.list_invitations — Method

list_invitations()
list_invitations(params::Dict{String,<:Any})

Lists all Security Hub membership invitations that were sent to the current AWS account. This operation is only used by accounts that are managed by invitation. Accounts that are managed using the integration with AWS Organizations do not receive invitations.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of items to return in the response.

"NextToken": The token that is required for pagination. On your first call to the ListInvitations operation, set the value of this parameter to NULL. For subsequent calls to the operation, to continue listing data, set the value of this parameter to the value returned from the previous response.

Main.Securityhub.list_members — Method

list_members()
list_members(params::Dict{String,<:Any})

Lists details about all member accounts for the current Security Hub administrator account. The results include both member accounts that belong to an organization and member accounts that were invited manually.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of items to return in the response.

"NextToken": The token that is required for pagination. On your first call to the ListMembers operation, set the value of this parameter to NULL. For subsequent calls to the operation, to continue listing data, set the value of this parameter to the value returned from the previous response.

"OnlyAssociated": Specifies which member accounts to include in the response based on their relationship status with the administrator account. The default value is TRUE. If OnlyAssociated is set to TRUE, the response includes member accounts whose relationship status with the administrator account is set to ENABLED. If OnlyAssociated is set to FALSE, the response includes all existing member accounts.

Main.Securityhub.list_organization_admin_accounts — Method

list_organization_admin_accounts()
list_organization_admin_accounts(params::Dict{String,<:Any})

Lists the Security Hub administrator accounts. Can only be called by the organization management account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of items to return in the response.

"NextToken": The token that is required for pagination. On your first call to the ListOrganizationAdminAccounts operation, set the value of this parameter to NULL. For subsequent calls to the operation, to continue listing data, set the value of this parameter to the value returned from the previous response.

Main.Securityhub.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Returns a list of tags associated with a resource.

Arguments

resource_arn: The ARN of the resource to retrieve tags for.

Main.Securityhub.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Adds one or more tags to a resource.

Arguments

resource_arn: The ARN of the resource to apply the tags to.

tags: The tags to add to the resource.

Main.Securityhub.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes one or more tags from a resource.

Arguments

resource_arn: The ARN of the resource to remove the tags from.

tag_keys: The tag keys associated with the tags to remove from the resource.

Main.Securityhub.update_action_target — Method

update_action_target(action_target_arn)
update_action_target(action_target_arn, params::Dict{String,<:Any})

Updates the name and description of a custom action target in Security Hub.

Arguments

action_target_arn: The ARN of the custom action target to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": The updated description for the custom action target.

"Name": The updated name of the custom action target.

Main.Securityhub.update_findings — Method

update_findings(filters)
update_findings(filters, params::Dict{String,<:Any})

UpdateFindings is deprecated. Instead of UpdateFindings, use BatchUpdateFindings. Updates the Note and RecordState of the Security Hub-aggregated findings that the filter attributes specify. Any member account that can view the finding also sees the update to the finding.

Arguments

filters: A collection of attributes that specify which findings you want to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Note": The updated note for the finding.

"RecordState": The updated record state for the finding.

Main.Securityhub.update_insight — Method

update_insight(insight_arn)
update_insight(insight_arn, params::Dict{String,<:Any})

Updates the Security Hub insight identified by the specified insight ARN.

Arguments

insight_arn: The ARN of the insight that you want to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Filters": The updated filters that define this insight.

"GroupByAttribute": The updated GroupBy attribute that defines this insight.

"Name": The updated name for the insight.

Main.Securityhub.update_organization_configuration — Method

update_organization_configuration(auto_enable)
update_organization_configuration(auto_enable, params::Dict{String,<:Any})

Used to update the configuration related to Organizations. Can only be called from a Security Hub administrator account.

Arguments

auto_enable: Whether to automatically enable Security Hub for new accounts in the organization. By default, this is false, and new accounts are not added automatically. To automatically enable Security Hub for new accounts, set this to true.

Main.Securityhub.update_security_hub_configuration — Method

update_security_hub_configuration()
update_security_hub_configuration(params::Dict{String,<:Any})

Updates configuration options for Security Hub.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AutoEnableControls": Whether to automatically enable new controls when they are added to standards that are enabled. By default, this is set to true, and new controls are enabled automatically. To not automatically enable new controls, set this to false.

Main.Securityhub.update_standards_control — Method

update_standards_control(standards_control_arn)
update_standards_control(standards_control_arn, params::Dict{String,<:Any})

Used to control whether an individual security standard control is enabled or disabled.

Arguments

standards_control_arn: The ARN of the security standard control to enable or disable.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ControlStatus": The updated status of the security standard control.

"DisabledReason": A description of the reason why you are disabling a security standard control. If you are disabling a control, then this is required.