Serverlessapplicationrepository

This page documents function available when using the Serverlessapplicationrepository module, created with @service Serverlessapplicationrepository.

Index

Main.Serverlessapplicationrepository.create_application

Main.Serverlessapplicationrepository.create_application_version

Main.Serverlessapplicationrepository.create_cloud_formation_change_set

Main.Serverlessapplicationrepository.create_cloud_formation_template

Main.Serverlessapplicationrepository.delete_application

Main.Serverlessapplicationrepository.get_application

Main.Serverlessapplicationrepository.get_application_policy

Main.Serverlessapplicationrepository.get_cloud_formation_template

Main.Serverlessapplicationrepository.list_application_dependencies

Main.Serverlessapplicationrepository.list_application_versions

Main.Serverlessapplicationrepository.list_applications

Main.Serverlessapplicationrepository.put_application_policy

Main.Serverlessapplicationrepository.unshare_application

Main.Serverlessapplicationrepository.update_application

Documentation

Main.Serverlessapplicationrepository.create_application — Method

CreateApplication()

Creates an application, optionally including an AWS SAM file to create the first application version in the same call.

Required Parameters

author: The name of the author publishing the app.Minimum length=1. Maximum length=127.Pattern "^a-z0-9?";

description: The description of the application.Minimum length=1. Maximum length=256

name: The name of the application that you want to publish.Minimum length=1. Maximum length=140Pattern: "[a-zA-Z0-9-]+";

Optional Parameters

homePageUrl: A URL with more information about the application, for example the location of your GitHub repository for the application.

labels: Labels to improve discovery of apps in search results.Minimum length=1. Maximum length=127. Maximum number of labels: 10Pattern: "^[a-zA-Z0-9+-_:/@]+";

licenseBody: A local text file that contains the license of the app that matches the spdxLicenseID value of your application.

The file has the format file://&lt;path>/&lt;filename>.Maximum size 5 MBYou can specify only one of licenseBody and licenseUrl; otherwise, an error results.

licenseUrl: A link to the S3 object that contains the license of the app that matches the spdxLicenseID value of your application.Maximum size 5 MBYou can specify only one of licenseBody and licenseUrl; otherwise, an error results.

readmeBody: A local text readme file in Markdown language that contains a more detailed description of the application and how it works.

The file has the format file://&lt;path>/&lt;filename>.Maximum size 5 MBYou can specify only one of readmeBody and readmeUrl; otherwise, an error results.

readmeUrl: A link to the S3 object in Markdown language that contains a more detailed description of the application and how it works.Maximum size 5 MBYou can specify only one of readmeBody and readmeUrl; otherwise, an error results.

semanticVersion: The semantic version of the application:

https://semver.org/

sourceCodeArchiveUrl: A link to the S3 object that contains the ZIP archive of the source code for this version of your application.Maximum size 50 MB

sourceCodeUrl: A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

spdxLicenseId: A valid identifier from https://spdx.org/licenses/.

templateBody: The local raw packaged AWS SAM template file of your application.

The file has the format file://&lt;path>/&lt;filename>.You can specify only one of templateBody and templateUrl; otherwise an error results.

templateUrl: A link to the S3 object containing the packaged AWS SAM template of your application.You can specify only one of templateBody and templateUrl; otherwise an error results.

Main.Serverlessapplicationrepository.create_application_version — Method

CreateApplicationVersion()

Creates an application version.

Required Parameters

applicationId: The Amazon Resource Name (ARN) of the application.

semanticVersion: The semantic version of the new version.

Optional Parameters

sourceCodeArchiveUrl: A link to the S3 object that contains the ZIP archive of the source code for this version of your application.Maximum size 50 MB

sourceCodeUrl: A link to a public repository for the source code of your application, for example the URL of a specific GitHub commit.

templateBody: The raw packaged AWS SAM template of your application.

templateUrl: A link to the packaged AWS SAM template of your application.

Main.Serverlessapplicationrepository.create_cloud_formation_change_set — Method

CreateCloudFormationChangeSet()

Creates an AWS CloudFormation change set for the given application.

Required Parameters

applicationId: The Amazon Resource Name (ARN) of the application.

stackName: This property corresponds to the parameter of the same name for the AWS CloudFormation CreateChangeSet API.

Optional Parameters

capabilities: A list of values that you must specify before you can deploy certain applications.

Some applications might include resources that can affect permissions in your AWS account, for example, by creating new AWS Identity and Access Management (IAM) users. For those applications, you must explicitly acknowledge their capabilities by specifying this parameter.The only valid values are CAPABILITYIAM, CAPABILITYNAMEDIAM, CAPABILITYRESOURCEPOLICY, and CAPABILITYAUTOEXPAND.The following resources require you to specify CAPABILITYIAM or CAPABILITYNAMEDIAM: AWS::IAM::Group, AWS::IAM::InstanceProfile, AWS::IAM::Policy, and AWS::IAM::Role. If the application contains IAM resources, you can specify either CAPABILITYIAM or CAPABILITYNAMEDIAM. If the application contains IAM resources with custom names, you must specify CAPABILITYNAMEDIAM.The following resources require you to specify CAPABILITYRESOURCEPOLICY: AWS::Lambda::Permission, AWS::IAM:Policy, AWS::ApplicationAutoScaling::ScalingPolicy, AWS::S3::BucketPolicy, AWS::SQS::QueuePolicy, and AWS::SNS:TopicPolicy.Applications that contain one or more nested applications require you to specify CAPABILITYAUTO_EXPAND.If your application template contains any of the above resources, we recommend that you review all permissions associated with the application before deploying. If you don't specify this parameter for an application that requires capabilities, the call will fail.

changeSetName: This property corresponds to the parameter of the same name for the AWS CloudFormation CreateChangeSet API.

clientToken: This property corresponds to the parameter of the same name for the AWS CloudFormation CreateChangeSet API.

description: This property corresponds to the parameter of the same name for the AWS CloudFormation CreateChangeSet API.

notificationArns: This property corresponds to the parameter of the same name for the AWS CloudFormation CreateChangeSet API.

parameterOverrides: A list of parameter values for the parameters of the application.

resourceTypes: This property corresponds to the parameter of the same name for the AWS CloudFormation CreateChangeSet API.

rollbackConfiguration: This property corresponds to the parameter of the same name for the AWS CloudFormation CreateChangeSet API.

semanticVersion: The semantic version of the application:

https://semver.org/

tags: This property corresponds to the parameter of the same name for the AWS CloudFormation CreateChangeSet API.

templateId: The UUID returned by CreateCloudFormationTemplate.Pattern: [0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}

Main.Serverlessapplicationrepository.create_cloud_formation_template — Method

CreateCloudFormationTemplate()

Creates an AWS CloudFormation template.

Required Parameters

applicationId: The Amazon Resource Name (ARN) of the application.

Optional Parameters

semanticVersion: The semantic version of the application:

https://semver.org/

Main.Serverlessapplicationrepository.delete_application — Method

DeleteApplication()

Deletes the specified application.

Required Parameters

applicationId: The Amazon Resource Name (ARN) of the application.

Main.Serverlessapplicationrepository.get_application — Method

GetApplication()

Gets the specified application.

Required Parameters

applicationId: The Amazon Resource Name (ARN) of the application.

Optional Parameters

semanticVersion: The semantic version of the application to get.

Main.Serverlessapplicationrepository.get_application_policy — Method

GetApplicationPolicy()

Retrieves the policy for the application.

Required Parameters

applicationId: The Amazon Resource Name (ARN) of the application.

Main.Serverlessapplicationrepository.get_cloud_formation_template — Method

GetCloudFormationTemplate()

Gets the specified AWS CloudFormation template.

Required Parameters

applicationId: The Amazon Resource Name (ARN) of the application.

templateId: The UUID returned by CreateCloudFormationTemplate.Pattern: [0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}

Main.Serverlessapplicationrepository.list_application_dependencies — Method

ListApplicationDependencies()

Retrieves the list of applications nested in the containing application.

Required Parameters

applicationId: The Amazon Resource Name (ARN) of the application.

Optional Parameters

maxItems: The total number of items to return.

nextToken: A token to specify where to start paginating.

semanticVersion: The semantic version of the application to get.

Main.Serverlessapplicationrepository.list_application_versions — Method

ListApplicationVersions()

Lists versions for the specified application.

Required Parameters

applicationId: The Amazon Resource Name (ARN) of the application.

Optional Parameters

maxItems: The total number of items to return.

nextToken: A token to specify where to start paginating.

Main.Serverlessapplicationrepository.list_applications — Method

ListApplications()

Lists applications owned by the requester.

Optional Parameters

maxItems: The total number of items to return.

nextToken: A token to specify where to start paginating.

Main.Serverlessapplicationrepository.put_application_policy — Method

PutApplicationPolicy()

Sets the permission policy for an application. For the list of actions supported for this operation, see Application Permissions .

Required Parameters

applicationId: The Amazon Resource Name (ARN) of the application.

statements: An array of policy statements applied to the application.

Main.Serverlessapplicationrepository.unshare_application — Method

UnshareApplication()

Unshares an application from an AWS Organization.This operation can be called only from the organization's master account.

Required Parameters

applicationId: The Amazon Resource Name (ARN) of the application.

organizationId: The AWS Organization ID to unshare the application from.

Main.Serverlessapplicationrepository.update_application — Method

UpdateApplication()

Updates the specified application.

Required Parameters

applicationId: The Amazon Resource Name (ARN) of the application.

Optional Parameters

author: The name of the author publishing the app.Minimum length=1. Maximum length=127.Pattern "^a-z0-9?";

description: The description of the application.Minimum length=1. Maximum length=256

homePageUrl: A URL with more information about the application, for example the location of your GitHub repository for the application.

labels: Labels to improve discovery of apps in search results.Minimum length=1. Maximum length=127. Maximum number of labels: 10Pattern: "^[a-zA-Z0-9+-_:/@]+";

readmeBody: A text readme file in Markdown language that contains a more detailed description of the application and how it works.Maximum size 5 MB

readmeUrl: A link to the readme file in Markdown language that contains a more detailed description of the application and how it works.Maximum size 5 MB