Sms

This page documents function available when using the Sms module, created with @service Sms.

Index

Main.Sms.create_app

Main.Sms.create_replication_job

Main.Sms.delete_app

Main.Sms.delete_app_launch_configuration

Main.Sms.delete_app_replication_configuration

Main.Sms.delete_app_validation_configuration

Main.Sms.delete_replication_job

Main.Sms.delete_server_catalog

Main.Sms.disassociate_connector

Main.Sms.generate_change_set

Main.Sms.generate_template

Main.Sms.get_app

Main.Sms.get_app_launch_configuration

Main.Sms.get_app_replication_configuration

Main.Sms.get_app_validation_configuration

Main.Sms.get_app_validation_output

Main.Sms.get_connectors

Main.Sms.get_replication_jobs

Main.Sms.get_replication_runs

Main.Sms.get_servers

Main.Sms.import_app_catalog

Main.Sms.import_server_catalog

Main.Sms.launch_app

Main.Sms.list_apps

Main.Sms.notify_app_validation_output

Main.Sms.put_app_launch_configuration

Main.Sms.put_app_replication_configuration

Main.Sms.put_app_validation_configuration

Main.Sms.start_app_replication

Main.Sms.start_on_demand_app_replication

Main.Sms.start_on_demand_replication_run

Main.Sms.stop_app_replication

Main.Sms.terminate_app

Main.Sms.update_app

Main.Sms.update_replication_job

Documentation

Main.Sms.create_app — Method

create_app()
create_app(params::Dict{String,<:Any})

Creates an application. An application consists of one or more server groups. Each server group contain one or more servers.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of application creation.

"description": The description of the new application

"name": The name of the new application.

"roleName": The name of the service role in the customer's account to be used by AWS SMS.

"serverGroups": The server groups to include in the application.

"tags": The tags to be associated with the application.

Main.Sms.create_replication_job — Method

create_replication_job(seed_replication_time, server_id)
create_replication_job(seed_replication_time, server_id, params::Dict{String,<:Any})

Creates a replication job. The replication job schedules periodic replication runs to replicate your server to AWS. Each replication run creates an Amazon Machine Image (AMI).

Arguments

seed_replication_time: The seed replication time.

server_id: The ID of the server.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The description of the replication job.

"encrypted": Indicates whether the replication job produces encrypted AMIs.

"frequency": The time between consecutive replication runs, in hours.

"kmsKeyId": The ID of the KMS key for replication jobs that produce encrypted AMIs. This value can be any of the following: KMS key ID KMS key alias ARN referring to the KMS key ID ARN referring to the KMS key alias If encrypted is true but a KMS key ID is not specified, the customer's default KMS key for Amazon EBS is used.

"licenseType": The license type to be used for the AMI created by a successful replication run.

"numberOfRecentAmisToKeep": The maximum number of SMS-created AMIs to retain. The oldest is deleted after the maximum number is reached and a new AMI is created.

"roleName": The name of the IAM role to be used by the AWS SMS.

"runOnce": Indicates whether to run the replication job one time.

Main.Sms.delete_app — Method

delete_app()
delete_app(params::Dict{String,<:Any})

Deletes the specified application. Optionally deletes the launched stack associated with the application and all AWS SMS replication jobs for servers in the application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"appId": The ID of the application.

"forceStopAppReplication": Indicates whether to stop all replication jobs corresponding to the servers in the application while deleting the application.

"forceTerminateApp": Indicates whether to terminate the stack corresponding to the application while deleting the application.

Main.Sms.delete_app_launch_configuration — Method

delete_app_launch_configuration()
delete_app_launch_configuration(params::Dict{String,<:Any})

Deletes the launch configuration for the specified application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"appId": The ID of the application.

Main.Sms.delete_app_replication_configuration — Method

delete_app_replication_configuration()
delete_app_replication_configuration(params::Dict{String,<:Any})

Deletes the replication configuration for the specified application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"appId": The ID of the application.

Main.Sms.delete_app_validation_configuration — Method

delete_app_validation_configuration(app_id)
delete_app_validation_configuration(app_id, params::Dict{String,<:Any})

Deletes the validation configuration for the specified application.

Arguments

app_id: The ID of the application.

Main.Sms.delete_replication_job — Method

delete_replication_job(replication_job_id)
delete_replication_job(replication_job_id, params::Dict{String,<:Any})

Deletes the specified replication job. After you delete a replication job, there are no further replication runs. AWS deletes the contents of the Amazon S3 bucket used to store AWS SMS artifacts. The AMIs created by the replication runs are not deleted.

Arguments

replication_job_id: The ID of the replication job.

Main.Sms.delete_server_catalog — Method

delete_server_catalog()
delete_server_catalog(params::Dict{String,<:Any})

Deletes all servers from your server catalog.

Main.Sms.disassociate_connector — Method

disassociate_connector(connector_id)
disassociate_connector(connector_id, params::Dict{String,<:Any})

Disassociates the specified connector from AWS SMS. After you disassociate a connector, it is no longer available to support replication jobs.

Arguments

connector_id: The ID of the connector.

Main.Sms.generate_change_set — Method

generate_change_set()
generate_change_set(params::Dict{String,<:Any})

Generates a target change set for a currently launched stack and writes it to an Amazon S3 object in the customer’s Amazon S3 bucket.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"appId": The ID of the application associated with the change set.

"changesetFormat": The format for the change set.

Main.Sms.generate_template — Method

generate_template()
generate_template(params::Dict{String,<:Any})

Generates an AWS CloudFormation template based on the current launch configuration and writes it to an Amazon S3 object in the customer’s Amazon S3 bucket.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"appId": The ID of the application associated with the AWS CloudFormation template.

"templateFormat": The format for generating the AWS CloudFormation template.

Main.Sms.get_app — Method

get_app()
get_app(params::Dict{String,<:Any})

Retrieve information about the specified application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"appId": The ID of the application.

Main.Sms.get_app_launch_configuration — Method

get_app_launch_configuration()
get_app_launch_configuration(params::Dict{String,<:Any})

Retrieves the application launch configuration associated with the specified application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"appId": The ID of the application.

Main.Sms.get_app_replication_configuration — Method

get_app_replication_configuration()
get_app_replication_configuration(params::Dict{String,<:Any})

Retrieves the application replication configuration associated with the specified application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"appId": The ID of the application.

Main.Sms.get_app_validation_configuration — Method

get_app_validation_configuration(app_id)
get_app_validation_configuration(app_id, params::Dict{String,<:Any})

Retrieves information about a configuration for validating an application.

Arguments

app_id: The ID of the application.

Main.Sms.get_app_validation_output — Method

get_app_validation_output(app_id)
get_app_validation_output(app_id, params::Dict{String,<:Any})

Retrieves output from validating an application.

Arguments

app_id: The ID of the application.

Main.Sms.get_connectors — Method

get_connectors()
get_connectors(params::Dict{String,<:Any})

Describes the connectors registered with the AWS SMS.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return in a single call. The default value is 50. To retrieve the remaining results, make another call with the returned NextToken value.

"nextToken": The token for the next set of results.

Main.Sms.get_replication_jobs — Method

get_replication_jobs()
get_replication_jobs(params::Dict{String,<:Any})

Describes the specified replication job or all of your replication jobs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return in a single call. The default value is 50. To retrieve the remaining results, make another call with the returned NextToken value.

"nextToken": The token for the next set of results.

"replicationJobId": The ID of the replication job.

Main.Sms.get_replication_runs — Method

get_replication_runs(replication_job_id)
get_replication_runs(replication_job_id, params::Dict{String,<:Any})

Describes the replication runs for the specified replication job.

Arguments

replication_job_id: The ID of the replication job.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return in a single call. The default value is 50. To retrieve the remaining results, make another call with the returned NextToken value.

"nextToken": The token for the next set of results.

Main.Sms.get_servers — Method

get_servers()
get_servers(params::Dict{String,<:Any})

Describes the servers in your server catalog. Before you can describe your servers, you must import them using ImportServerCatalog.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return in a single call. The default value is 50. To retrieve the remaining results, make another call with the returned NextToken value.

"nextToken": The token for the next set of results.

"vmServerAddressList": The server addresses.

Main.Sms.import_app_catalog — Method

import_app_catalog()
import_app_catalog(params::Dict{String,<:Any})

Allows application import from AWS Migration Hub.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"roleName": The name of the service role. If you omit this parameter, we create a service-linked role for AWS Migration Hub in your account. Otherwise, the role that you provide must have the policy and trust policy described in the AWS Migration Hub User Guide.

Main.Sms.import_server_catalog — Method

import_server_catalog()
import_server_catalog(params::Dict{String,<:Any})

Gathers a complete list of on-premises servers. Connectors must be installed and monitoring all servers to import. This call returns immediately, but might take additional time to retrieve all the servers.

Main.Sms.launch_app — Method

launch_app()
launch_app(params::Dict{String,<:Any})

Launches the specified application as a stack in AWS CloudFormation.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"appId": The ID of the application.

Main.Sms.list_apps — Method

list_apps()
list_apps(params::Dict{String,<:Any})

Retrieves summaries for all applications.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"appIds": The unique application IDs.

"maxResults": The maximum number of results to return in a single call. The default value is 100. To retrieve the remaining results, make another call with the returned NextToken value.

"nextToken": The token for the next set of results.

Main.Sms.notify_app_validation_output — Method

notify_app_validation_output(app_id)
notify_app_validation_output(app_id, params::Dict{String,<:Any})

Provides information to AWS SMS about whether application validation is successful.

Arguments

app_id: The ID of the application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"notificationContext": The notification information.

Main.Sms.put_app_launch_configuration — Method

put_app_launch_configuration()
put_app_launch_configuration(params::Dict{String,<:Any})

Creates or updates the launch configuration for the specified application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"appId": The ID of the application.

"autoLaunch": Indicates whether the application is configured to launch automatically after replication is complete.

"roleName": The name of service role in the customer's account that AWS CloudFormation uses to launch the application.

"serverGroupLaunchConfigurations": Information about the launch configurations for server groups in the application.

Main.Sms.put_app_replication_configuration — Method

put_app_replication_configuration()
put_app_replication_configuration(params::Dict{String,<:Any})

Creates or updates the replication configuration for the specified application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"appId": The ID of the application.

"serverGroupReplicationConfigurations": Information about the replication configurations for server groups in the application.

Main.Sms.put_app_validation_configuration — Method

put_app_validation_configuration(app_id)
put_app_validation_configuration(app_id, params::Dict{String,<:Any})

Creates or updates a validation configuration for the specified application.

Arguments

app_id: The ID of the application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"appValidationConfigurations": The configuration for application validation.

"serverGroupValidationConfigurations": The configuration for instance validation.

Main.Sms.start_app_replication — Method

start_app_replication()
start_app_replication(params::Dict{String,<:Any})

Starts replicating the specified application by creating replication jobs for each server in the application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"appId": The ID of the application.

Main.Sms.start_on_demand_app_replication — Method

start_on_demand_app_replication(app_id)
start_on_demand_app_replication(app_id, params::Dict{String,<:Any})

Starts an on-demand replication run for the specified application.

Arguments

app_id: The ID of the application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The description of the replication run.

Main.Sms.start_on_demand_replication_run — Method

start_on_demand_replication_run(replication_job_id)
start_on_demand_replication_run(replication_job_id, params::Dict{String,<:Any})

Starts an on-demand replication run for the specified replication job. This replication run starts immediately. This replication run is in addition to the ones already scheduled. There is a limit on the number of on-demand replications runs that you can request in a 24-hour period.

Arguments

replication_job_id: The ID of the replication job.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The description of the replication run.

Main.Sms.stop_app_replication — Method

stop_app_replication()
stop_app_replication(params::Dict{String,<:Any})

Stops replicating the specified application by deleting the replication job for each server in the application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"appId": The ID of the application.

Main.Sms.terminate_app — Method

terminate_app()
terminate_app(params::Dict{String,<:Any})

Terminates the stack for the specified application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"appId": The ID of the application.

Main.Sms.update_app — Method

update_app()
update_app(params::Dict{String,<:Any})

Updates the specified application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"appId": The ID of the application.

"description": The new description of the application.

"name": The new name of the application.

"roleName": The name of the service role in the customer's account used by AWS SMS.

"serverGroups": The server groups in the application to update.

"tags": The tags to associate with the application.

Main.Sms.update_replication_job — Method

update_replication_job(replication_job_id)
update_replication_job(replication_job_id, params::Dict{String,<:Any})

Updates the specified settings for the specified replication job.

Arguments

replication_job_id: The ID of the replication job.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The description of the replication job.

"encrypted": When true, the replication job produces encrypted AMIs. For more information, KmsKeyId.

"frequency": The time between consecutive replication runs, in hours.

"kmsKeyId": The ID of the KMS key for replication jobs that produce encrypted AMIs. This value can be any of the following: KMS key ID KMS key alias ARN referring to the KMS key ID ARN referring to the KMS key alias If encrypted is enabled but a KMS key ID is not specified, the customer's default KMS key for Amazon EBS is used.

"licenseType": The license type to be used for the AMI created by a successful replication run.

"nextReplicationRunStartTime": The start time of the next replication run.

"numberOfRecentAmisToKeep": The maximum number of SMS-created AMIs to retain. The oldest is deleted after the maximum number is reached and a new AMI is created.

"roleName": The name of the IAM role to be used by AWS SMS.