Socialmessaging

This page documents function available when using the Socialmessaging module, created with @service Socialmessaging.

Index

Main.Socialmessaging.associate_whats_app_business_account

Main.Socialmessaging.create_whats_app_message_template

Main.Socialmessaging.create_whats_app_message_template_from_library

Main.Socialmessaging.create_whats_app_message_template_media

Main.Socialmessaging.delete_whats_app_message_media

Main.Socialmessaging.delete_whats_app_message_template

Main.Socialmessaging.disassociate_whats_app_business_account

Main.Socialmessaging.get_linked_whats_app_business_account

Main.Socialmessaging.get_linked_whats_app_business_account_phone_number

Main.Socialmessaging.get_whats_app_message_media

Main.Socialmessaging.get_whats_app_message_template

Main.Socialmessaging.list_linked_whats_app_business_accounts

Main.Socialmessaging.list_tags_for_resource

Main.Socialmessaging.list_whats_app_message_templates

Main.Socialmessaging.list_whats_app_template_library

Main.Socialmessaging.post_whats_app_message_media

Main.Socialmessaging.put_whats_app_business_account_event_destinations

Main.Socialmessaging.send_whats_app_message

Main.Socialmessaging.tag_resource

Main.Socialmessaging.untag_resource

Main.Socialmessaging.update_whats_app_message_template

Documentation

Main.Socialmessaging.associate_whats_app_business_account — Function

associate_whats_app_business_account()
associate_whats_app_business_account(params::Dict{String,<:Any})

This is only used through the Amazon Web Services console during sign-up to associate your WhatsApp Business Account to your Amazon Web Services account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"setupFinalization": A JSON object that contains the phone numbers and WhatsApp Business Account to link to your account.

"signupCallback": Contains the callback access token.

Main.Socialmessaging.create_whats_app_message_template — Function

create_whats_app_message_template(id, template_definition)
create_whats_app_message_template(id, template_definition, params::Dict{String,<:Any})

Creates a new WhatsApp message template from a custom definition.

Note

Amazon Web Services End User Messaging Social does not store any WhatsApp message template content.

Arguments

id: The ID of the WhatsApp Business Account to associate with this template.

template_definition: The complete template definition as a JSON blob.

Main.Socialmessaging.create_whats_app_message_template_from_library — Function

create_whats_app_message_template_from_library(id, meta_library_template)
create_whats_app_message_template_from_library(id, meta_library_template, params::Dict{String,<:Any})

Creates a new WhatsApp message template using a template from Meta's template library.

Arguments

id: The ID of the WhatsApp Business Account to associate with this template.

meta_library_template: The template configuration from Meta's library, including customizations for buttons and body text.

Main.Socialmessaging.create_whats_app_message_template_media — Function

create_whats_app_message_template_media(id)
create_whats_app_message_template_media(id, params::Dict{String,<:Any})

Uploads media for use in a WhatsApp message template.

Arguments

id: The ID of the WhatsApp Business Account associated with this media upload.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"sourceS3File":

Main.Socialmessaging.delete_whats_app_message_media — Function

delete_whats_app_message_media(media_id, origination_phone_number_id)
delete_whats_app_message_media(media_id, origination_phone_number_id, params::Dict{String,<:Any})

Delete a media object from the WhatsApp service. If the object is still in an Amazon S3 bucket you should delete it from there too.

Arguments

media_id: The unique identifier of the media file to delete. Use the mediaId returned from PostWhatsAppMessageMedia.

origination_phone_number_id: The unique identifier of the originating phone number associated with the media. Phone number identifiers are formatted as phone-number-id-01234567890123456789012345678901. Use GetLinkedWhatsAppBusinessAccount to find a phone number's id.

Main.Socialmessaging.delete_whats_app_message_template — Function

delete_whats_app_message_template(id, template_name)
delete_whats_app_message_template(id, template_name, params::Dict{String,<:Any})

Deletes a WhatsApp message template.

Arguments

id: The ID of the WhatsApp Business Account associated with this template.

template_name: The name of the template to delete.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"deleteAllTemplates": If true, deletes all language versions of the template.

"metaTemplateId": The numeric ID of the template assigned by Meta.

Main.Socialmessaging.disassociate_whats_app_business_account — Function

disassociate_whats_app_business_account(id)
disassociate_whats_app_business_account(id, params::Dict{String,<:Any})

Disassociate a WhatsApp Business Account (WABA) from your Amazon Web Services account.

Arguments

id: The unique identifier of your WhatsApp Business Account. WABA identifiers are formatted as waba-01234567890123456789012345678901. Use ListLinkedWhatsAppBusinessAccounts to list all WABAs and their details.

Main.Socialmessaging.get_linked_whats_app_business_account — Function

get_linked_whats_app_business_account(id)
get_linked_whats_app_business_account(id, params::Dict{String,<:Any})

Get the details of your linked WhatsApp Business Account.

Arguments

id: The unique identifier, from Amazon Web Services, of the linked WhatsApp Business Account. WABA identifiers are formatted as waba-01234567890123456789012345678901. Use ListLinkedWhatsAppBusinessAccounts to list all WABAs and their details.

Main.Socialmessaging.get_linked_whats_app_business_account_phone_number — Function

get_linked_whats_app_business_account_phone_number(id)
get_linked_whats_app_business_account_phone_number(id, params::Dict{String,<:Any})

Retrieve the WABA account id and phone number details of a WhatsApp business account phone number.

Arguments

id: The unique identifier of the phone number. Phone number identifiers are formatted as phone-number-id-01234567890123456789012345678901. Use GetLinkedWhatsAppBusinessAccount to find a phone number's id.

Main.Socialmessaging.get_whats_app_message_media — Function

get_whats_app_message_media(media_id, origination_phone_number_id)
get_whats_app_message_media(media_id, origination_phone_number_id, params::Dict{String,<:Any})

Get a media file from the WhatsApp service. On successful completion the media file is retrieved from Meta and stored in the specified Amazon S3 bucket. Use either destinationS3File or destinationS3PresignedUrl for the destination. If both are used then an InvalidParameterException is returned.

Arguments

media_id: The unique identifier for the media file.

origination_phone_number_id: The unique identifier of the originating phone number for the WhatsApp message media. The phone number identifiers are formatted as phone-number-id-01234567890123456789012345678901. Use GetLinkedWhatsAppBusinessAccount to find a phone number's id.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"destinationS3File": The bucketName and key of the S3 media file.

"destinationS3PresignedUrl": The presign url of the media file.

"metadataOnly": Set to True to get only the metadata for the file.

Main.Socialmessaging.get_whats_app_message_template — Function

get_whats_app_message_template(id, meta_template_id)
get_whats_app_message_template(id, meta_template_id, params::Dict{String,<:Any})

Retrieves a specific WhatsApp message template.

Arguments

id: The ID of the WhatsApp Business Account associated with this template.

meta_template_id: The numeric ID of the template assigned by Meta.

Main.Socialmessaging.list_linked_whats_app_business_accounts — Function

list_linked_whats_app_business_accounts()
list_linked_whats_app_business_accounts(params::Dict{String,<:Any})

List all WhatsApp Business Accounts linked to your Amazon Web Services account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return.

"nextToken": The next token for pagination.

Main.Socialmessaging.list_tags_for_resource — Function

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

List all tags associated with a resource, such as a phone number or WABA.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource to retrieve the tags from.

Main.Socialmessaging.list_whats_app_message_templates — Function

list_whats_app_message_templates(id)
list_whats_app_message_templates(id, params::Dict{String,<:Any})

Lists WhatsApp message templates for a specific WhatsApp Business Account.

Arguments

id: The ID of the WhatsApp Business Account to list templates for.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return per page (1-100).

"nextToken": The token for the next page of results.

Main.Socialmessaging.list_whats_app_template_library — Function

list_whats_app_template_library(id)
list_whats_app_template_library(id, params::Dict{String,<:Any})

Lists templates available in Meta's template library for WhatsApp messaging.

Arguments

id: The ID of the WhatsApp Business Account to list library templates for.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filters": Map of filters to apply (searchKey, topic, usecase, industry, language).

"maxResults": The maximum number of results to return per page (1-100).

"nextToken": The token for the next page of results.

Main.Socialmessaging.post_whats_app_message_media — Function

post_whats_app_message_media(origination_phone_number_id)
post_whats_app_message_media(origination_phone_number_id, params::Dict{String,<:Any})

Upload a media file to the WhatsApp service. Only the specified originationPhoneNumberId has the permissions to send the media file when using SendWhatsAppMessage. You must use either sourceS3File or sourceS3PresignedUrl for the source. If both or neither are specified then an InvalidParameterException is returned.

Arguments

origination_phone_number_id: The ID of the phone number to associate with the WhatsApp media file. The phone number identifiers are formatted as phone-number-id-01234567890123456789012345678901. Use GetLinkedWhatsAppBusinessAccount to find a phone number's id.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"sourceS3File": The source S3 url for the media file.

"sourceS3PresignedUrl": The source presign url of the media file.

Main.Socialmessaging.put_whats_app_business_account_event_destinations — Function

put_whats_app_business_account_event_destinations(event_destinations, id)
put_whats_app_business_account_event_destinations(event_destinations, id, params::Dict{String,<:Any})

Add an event destination to log event data from WhatsApp for a WhatsApp Business Account (WABA). A WABA can only have one event destination at a time. All resources associated with the WABA use the same event destination.

Arguments

event_destinations: An array of WhatsAppBusinessAccountEventDestination event destinations.

id: The unique identifier of your WhatsApp Business Account. WABA identifiers are formatted as waba-01234567890123456789012345678901. Use ListLinkedWhatsAppBusinessAccounts to list all WABAs and their details.

Main.Socialmessaging.send_whats_app_message — Function

send_whats_app_message(message, meta_api_version, origination_phone_number_id)
send_whats_app_message(message, meta_api_version, origination_phone_number_id, params::Dict{String,<:Any})

Send a WhatsApp message. For examples of sending a message using the Amazon Web Services CLI, see Sending messages in the <i>Amazon Web Services End User Messaging Social User Guide </i>.

Arguments

message: The message to send through WhatsApp. The length is in KB. The message field passes through a WhatsApp Message object, see Messages in the WhatsApp Business Platform Cloud API Reference.

meta_api_version: The API version for the request formatted as v{VersionNumber}. For a list of supported API versions and Amazon Web Services Regions, see Amazon Web Services End User Messaging Social API Service Endpoints in the Amazon Web Services General Reference.

origination_phone_number_id: The ID of the phone number used to send the WhatsApp message. If you are sending a media file only the originationPhoneNumberId used to upload the file can be used. Phone number identifiers are formatted as phone-number-id-01234567890123456789012345678901. Use GetLinkedWhatsAppBusinessAccount to find a phone number's id.

Main.Socialmessaging.tag_resource — Function

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Adds or overwrites only the specified tags for the specified resource. When you specify an existing tag key, the value is overwritten with the new value.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource to tag.

tags: The tags to add to the resource.

Main.Socialmessaging.untag_resource — Function

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes the specified tags from a resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource to remove tags from.

tag_keys: The keys of the tags to remove from the resource.

Main.Socialmessaging.update_whats_app_message_template — Function

update_whats_app_message_template(id, meta_template_id)
update_whats_app_message_template(id, meta_template_id, params::Dict{String,<:Any})

Updates an existing WhatsApp message template.

Arguments

id: The ID of the WhatsApp Business Account associated with this template.

meta_template_id: The numeric ID of the template assigned by Meta.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ctaUrlLinkTrackingOptedOut": When true, disables click tracking for call-to-action URL buttons in the template.

"parameterFormat": The format specification for parameters in the template, this can be either 'named' or 'positional'.

"templateCategory": The new category for the template (for example, UTILITY or MARKETING).

"templateComponents": The updated components of the template as a JSON blob (maximum 3000 characters).