Sso Admin

This page documents function available when using the Sso_Admin module, created with @service Sso_Admin.

Index

Main.Sso_Admin.attach_customer_managed_policy_reference_to_permission_set

Main.Sso_Admin.attach_managed_policy_to_permission_set

Main.Sso_Admin.create_account_assignment

Main.Sso_Admin.create_instance_access_control_attribute_configuration

Main.Sso_Admin.create_permission_set

Main.Sso_Admin.delete_account_assignment

Main.Sso_Admin.delete_inline_policy_from_permission_set

Main.Sso_Admin.delete_instance_access_control_attribute_configuration

Main.Sso_Admin.delete_permission_set

Main.Sso_Admin.delete_permissions_boundary_from_permission_set

Main.Sso_Admin.describe_account_assignment_creation_status

Main.Sso_Admin.describe_account_assignment_deletion_status

Main.Sso_Admin.describe_instance_access_control_attribute_configuration

Main.Sso_Admin.describe_permission_set

Main.Sso_Admin.describe_permission_set_provisioning_status

Main.Sso_Admin.detach_customer_managed_policy_reference_from_permission_set

Main.Sso_Admin.detach_managed_policy_from_permission_set

Main.Sso_Admin.get_inline_policy_for_permission_set

Main.Sso_Admin.get_permissions_boundary_for_permission_set

Main.Sso_Admin.list_account_assignment_creation_status

Main.Sso_Admin.list_account_assignment_deletion_status

Main.Sso_Admin.list_account_assignments

Main.Sso_Admin.list_accounts_for_provisioned_permission_set

Main.Sso_Admin.list_customer_managed_policy_references_in_permission_set

Main.Sso_Admin.list_instances

Main.Sso_Admin.list_managed_policies_in_permission_set

Main.Sso_Admin.list_permission_set_provisioning_status

Main.Sso_Admin.list_permission_sets

Main.Sso_Admin.list_permission_sets_provisioned_to_account

Main.Sso_Admin.list_tags_for_resource

Main.Sso_Admin.provision_permission_set

Main.Sso_Admin.put_inline_policy_to_permission_set

Main.Sso_Admin.put_permissions_boundary_to_permission_set

Main.Sso_Admin.tag_resource

Main.Sso_Admin.untag_resource

Main.Sso_Admin.update_instance_access_control_attribute_configuration

Main.Sso_Admin.update_permission_set

Documentation

Main.Sso_Admin.attach_customer_managed_policy_reference_to_permission_set — Method

attach_customer_managed_policy_reference_to_permission_set(customer_managed_policy_reference, instance_arn, permission_set_arn)
attach_customer_managed_policy_reference_to_permission_set(customer_managed_policy_reference, instance_arn, permission_set_arn, params::Dict{String,<:Any})

Attaches the specified customer managed policy to the specified PermissionSet.

Arguments

customer_managed_policy_reference: Specifies the name and path of a customer managed policy. You must have an IAM policy that matches the name and path in each AWS account where you want to deploy your permission set.

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed.

permission_set_arn: The ARN of the PermissionSet.

Main.Sso_Admin.attach_managed_policy_to_permission_set — Method

attach_managed_policy_to_permission_set(instance_arn, managed_policy_arn, permission_set_arn)
attach_managed_policy_to_permission_set(instance_arn, managed_policy_arn, permission_set_arn, params::Dict{String,<:Any})

Attaches an AWS managed policy ARN to a permission set. If the permission set is already referenced by one or more account assignments, you will need to call ProvisionPermissionSet after this operation. Calling ProvisionPermissionSet applies the corresponding IAM policy updates to all assigned accounts.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

managed_policy_arn: The AWS managed policy ARN to be attached to a permission set.

permission_set_arn: The ARN of the PermissionSet that the managed policy should be attached to.

Main.Sso_Admin.create_account_assignment — Method

create_account_assignment(instance_arn, permission_set_arn, principal_id, principal_type, target_id, target_type)
create_account_assignment(instance_arn, permission_set_arn, principal_id, principal_type, target_id, target_type, params::Dict{String,<:Any})

Assigns access to a principal for a specified AWS account using a specified permission set. The term principal here refers to a user or group that is defined in IAM Identity Center. As part of a successful CreateAccountAssignment call, the specified permission set will automatically be provisioned to the account in the form of an IAM policy. That policy is attached to the IAM role created in IAM Identity Center. If the permission set is subsequently updated, the corresponding IAM policies attached to roles in your accounts will not be updated automatically. In this case, you must call ProvisionPermissionSet to make these updates. After a successful response, call DescribeAccountAssignmentCreationStatus to describe the status of an assignment creation request.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

permission_set_arn: The ARN of the permission set that the admin wants to grant the principal access to.

principal_id: An identifier for an object in IAM Identity Center, such as a user or group. PrincipalIds are GUIDs (For example, f81d4fae-7dec-11d0-a765-00a0c91e6bf6). For more information about PrincipalIds in IAM Identity Center, see the IAM Identity Center Identity Store API Reference.

principal_type: The entity type for which the assignment will be created.

target_id: TargetID is an AWS account identifier, typically a 10-12 digit string (For example, 123456789012).

target_type: The entity type for which the assignment will be created.

Main.Sso_Admin.create_instance_access_control_attribute_configuration — Method

create_instance_access_control_attribute_configuration(instance_access_control_attribute_configuration, instance_arn)
create_instance_access_control_attribute_configuration(instance_access_control_attribute_configuration, instance_arn, params::Dict{String,<:Any})

Enables the attributes-based access control (ABAC) feature for the specified IAM Identity Center instance. You can also specify new attributes to add to your ABAC configuration during the enabling process. For more information about ABAC, see Attribute-Based Access Control in the IAM Identity Center User Guide. After a successful response, call DescribeInstanceAccessControlAttributeConfiguration to validate that InstanceAccessControlAttributeConfiguration was created.

Arguments

instance_access_control_attribute_configuration: Specifies the IAM Identity Center identity store attributes to add to your ABAC configuration. When using an external identity provider as an identity source, you can pass attributes through the SAML assertion. Doing so provides an alternative to configuring attributes from the IAM Identity Center identity store. If a SAML assertion passes any of these attributes, IAM Identity Center will replace the attribute value with the value from the IAM Identity Center identity store.

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed.

Main.Sso_Admin.create_permission_set — Method

create_permission_set(instance_arn, name)
create_permission_set(instance_arn, name, params::Dict{String,<:Any})

Creates a permission set within a specified IAM Identity Center instance. To grant users and groups access to AWS account resources, use CreateAccountAssignment .

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

name: The name of the PermissionSet.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": The description of the PermissionSet.

"RelayState": Used to redirect users within the application during the federation authentication process.

"SessionDuration": The length of time that the application user sessions are valid in the ISO-8601 standard.

"Tags": The tags to attach to the new PermissionSet.

Main.Sso_Admin.delete_account_assignment — Method

delete_account_assignment(instance_arn, permission_set_arn, principal_id, principal_type, target_id, target_type)
delete_account_assignment(instance_arn, permission_set_arn, principal_id, principal_type, target_id, target_type, params::Dict{String,<:Any})

Deletes a principal's access from a specified AWS account using a specified permission set. After a successful response, call DescribeAccountAssignmentCreationStatus to describe the status of an assignment deletion request.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

permission_set_arn: The ARN of the permission set that will be used to remove access.

principal_id: An identifier for an object in IAM Identity Center, such as a user or group. PrincipalIds are GUIDs (For example, f81d4fae-7dec-11d0-a765-00a0c91e6bf6). For more information about PrincipalIds in IAM Identity Center, see the IAM Identity Center Identity Store API Reference.

principal_type: The entity type for which the assignment will be deleted.

target_id: TargetID is an AWS account identifier, typically a 10-12 digit string (For example, 123456789012).

target_type: The entity type for which the assignment will be deleted.

Main.Sso_Admin.delete_inline_policy_from_permission_set — Method

delete_inline_policy_from_permission_set(instance_arn, permission_set_arn)
delete_inline_policy_from_permission_set(instance_arn, permission_set_arn, params::Dict{String,<:Any})

Deletes the inline policy from a specified permission set.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

permission_set_arn: The ARN of the permission set that will be used to remove access.

Main.Sso_Admin.delete_instance_access_control_attribute_configuration — Method

delete_instance_access_control_attribute_configuration(instance_arn)
delete_instance_access_control_attribute_configuration(instance_arn, params::Dict{String,<:Any})

Disables the attributes-based access control (ABAC) feature for the specified IAM Identity Center instance and deletes all of the attribute mappings that have been configured. Once deleted, any attributes that are received from an identity source and any custom attributes you have previously configured will not be passed. For more information about ABAC, see Attribute-Based Access Control in the IAM Identity Center User Guide.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed.

Main.Sso_Admin.delete_permission_set — Method

delete_permission_set(instance_arn, permission_set_arn)
delete_permission_set(instance_arn, permission_set_arn, params::Dict{String,<:Any})

Deletes the specified permission set.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

permission_set_arn: The ARN of the permission set that should be deleted.

Main.Sso_Admin.delete_permissions_boundary_from_permission_set — Method

delete_permissions_boundary_from_permission_set(instance_arn, permission_set_arn)
delete_permissions_boundary_from_permission_set(instance_arn, permission_set_arn, params::Dict{String,<:Any})

Deletes the permissions boundary from a specified PermissionSet.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed.

permission_set_arn: The ARN of the PermissionSet.

Main.Sso_Admin.describe_account_assignment_creation_status — Method

describe_account_assignment_creation_status(account_assignment_creation_request_id, instance_arn)
describe_account_assignment_creation_status(account_assignment_creation_request_id, instance_arn, params::Dict{String,<:Any})

Describes the status of the assignment creation request.

Arguments

account_assignment_creation_request_id: The identifier that is used to track the request operation progress.

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

Main.Sso_Admin.describe_account_assignment_deletion_status — Method

describe_account_assignment_deletion_status(account_assignment_deletion_request_id, instance_arn)
describe_account_assignment_deletion_status(account_assignment_deletion_request_id, instance_arn, params::Dict{String,<:Any})

Describes the status of the assignment deletion request.

Arguments

account_assignment_deletion_request_id: The identifier that is used to track the request operation progress.

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

Main.Sso_Admin.describe_instance_access_control_attribute_configuration — Method

describe_instance_access_control_attribute_configuration(instance_arn)
describe_instance_access_control_attribute_configuration(instance_arn, params::Dict{String,<:Any})

Returns the list of IAM Identity Center identity store attributes that have been configured to work with attributes-based access control (ABAC) for the specified IAM Identity Center instance. This will not return attributes configured and sent by an external identity provider. For more information about ABAC, see Attribute-Based Access Control in the IAM Identity Center User Guide.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed.

Main.Sso_Admin.describe_permission_set — Method

describe_permission_set(instance_arn, permission_set_arn)
describe_permission_set(instance_arn, permission_set_arn, params::Dict{String,<:Any})

Gets the details of the permission set.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

permission_set_arn: The ARN of the permission set.

Main.Sso_Admin.describe_permission_set_provisioning_status — Method

describe_permission_set_provisioning_status(instance_arn, provision_permission_set_request_id)
describe_permission_set_provisioning_status(instance_arn, provision_permission_set_request_id, params::Dict{String,<:Any})

Describes the status for the given permission set provisioning request.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

provision_permission_set_request_id: The identifier that is provided by the ProvisionPermissionSet call to retrieve the current status of the provisioning workflow.

Main.Sso_Admin.detach_customer_managed_policy_reference_from_permission_set — Method

detach_customer_managed_policy_reference_from_permission_set(customer_managed_policy_reference, instance_arn, permission_set_arn)
detach_customer_managed_policy_reference_from_permission_set(customer_managed_policy_reference, instance_arn, permission_set_arn, params::Dict{String,<:Any})

Detaches the specified customer managed policy from the specified PermissionSet.

Arguments

customer_managed_policy_reference: Specifies the name and path of a customer managed policy. You must have an IAM policy that matches the name and path in each AWS account where you want to deploy your permission set.

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed.

permission_set_arn: The ARN of the PermissionSet.

Main.Sso_Admin.detach_managed_policy_from_permission_set — Method

detach_managed_policy_from_permission_set(instance_arn, managed_policy_arn, permission_set_arn)
detach_managed_policy_from_permission_set(instance_arn, managed_policy_arn, permission_set_arn, params::Dict{String,<:Any})

Detaches the attached AWS managed policy ARN from the specified permission set.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

managed_policy_arn: The AWS managed policy ARN to be detached from a permission set.

permission_set_arn: The ARN of the PermissionSet from which the policy should be detached.

Main.Sso_Admin.get_inline_policy_for_permission_set — Method

get_inline_policy_for_permission_set(instance_arn, permission_set_arn)
get_inline_policy_for_permission_set(instance_arn, permission_set_arn, params::Dict{String,<:Any})

Obtains the inline policy assigned to the permission set.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

permission_set_arn: The ARN of the permission set.

Main.Sso_Admin.get_permissions_boundary_for_permission_set — Method

get_permissions_boundary_for_permission_set(instance_arn, permission_set_arn)
get_permissions_boundary_for_permission_set(instance_arn, permission_set_arn, params::Dict{String,<:Any})

Obtains the permissions boundary for a specified PermissionSet.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed.

permission_set_arn: The ARN of the PermissionSet.

Main.Sso_Admin.list_account_assignment_creation_status — Method

list_account_assignment_creation_status(instance_arn)
list_account_assignment_creation_status(instance_arn, params::Dict{String,<:Any})

Lists the status of the AWS account assignment creation requests for a specified IAM Identity Center instance.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Filter": Filters results based on the passed attribute value.

"MaxResults": The maximum number of results to display for the assignment.

"NextToken": The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

Main.Sso_Admin.list_account_assignment_deletion_status — Method

list_account_assignment_deletion_status(instance_arn)
list_account_assignment_deletion_status(instance_arn, params::Dict{String,<:Any})

Lists the status of the AWS account assignment deletion requests for a specified IAM Identity Center instance.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Filter": Filters results based on the passed attribute value.

"MaxResults": The maximum number of results to display for the assignment.

"NextToken": The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

Main.Sso_Admin.list_account_assignments — Method

list_account_assignments(account_id, instance_arn, permission_set_arn)
list_account_assignments(account_id, instance_arn, permission_set_arn, params::Dict{String,<:Any})

Lists the assignee of the specified AWS account with the specified permission set.

Arguments

account_id: The identifier of the AWS account from which to list the assignments.

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

permission_set_arn: The ARN of the permission set from which to list assignments.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to display for the assignment.

"NextToken": The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

Main.Sso_Admin.list_accounts_for_provisioned_permission_set — Method

list_accounts_for_provisioned_permission_set(instance_arn, permission_set_arn)
list_accounts_for_provisioned_permission_set(instance_arn, permission_set_arn, params::Dict{String,<:Any})

Lists all the AWS accounts where the specified permission set is provisioned.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

permission_set_arn: The ARN of the PermissionSet from which the associated AWS accounts will be listed.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to display for the PermissionSet.

"NextToken": The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

"ProvisioningStatus": The permission set provisioning status for an AWS account.

Main.Sso_Admin.list_customer_managed_policy_references_in_permission_set — Method

list_customer_managed_policy_references_in_permission_set(instance_arn, permission_set_arn)
list_customer_managed_policy_references_in_permission_set(instance_arn, permission_set_arn, params::Dict{String,<:Any})

Lists all customer managed policies attached to a specified PermissionSet.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed.

permission_set_arn: The ARN of the PermissionSet.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to display for the list call.

"NextToken": The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

Main.Sso_Admin.list_instances — Method

list_instances()
list_instances(params::Dict{String,<:Any})

Lists the IAM Identity Center instances that the caller has access to.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to display for the instance.

"NextToken": The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

Main.Sso_Admin.list_managed_policies_in_permission_set — Method

list_managed_policies_in_permission_set(instance_arn, permission_set_arn)
list_managed_policies_in_permission_set(instance_arn, permission_set_arn, params::Dict{String,<:Any})

Lists the AWS managed policy that is attached to a specified permission set.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

permission_set_arn: The ARN of the PermissionSet whose managed policies will be listed.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to display for the PermissionSet.

"NextToken": The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

Main.Sso_Admin.list_permission_set_provisioning_status — Method

list_permission_set_provisioning_status(instance_arn)
list_permission_set_provisioning_status(instance_arn, params::Dict{String,<:Any})

Lists the status of the permission set provisioning requests for a specified IAM Identity Center instance.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Filter": Filters results based on the passed attribute value.

"MaxResults": The maximum number of results to display for the assignment.

"NextToken": The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

Main.Sso_Admin.list_permission_sets — Method

list_permission_sets(instance_arn)
list_permission_sets(instance_arn, params::Dict{String,<:Any})

Lists the PermissionSets in an IAM Identity Center instance.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to display for the assignment.

"NextToken": The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

Main.Sso_Admin.list_permission_sets_provisioned_to_account — Method

list_permission_sets_provisioned_to_account(account_id, instance_arn)
list_permission_sets_provisioned_to_account(account_id, instance_arn, params::Dict{String,<:Any})

Lists all the permission sets that are provisioned to a specified AWS account.

Arguments

account_id: The identifier of the AWS account from which to list the assignments.

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to display for the assignment.

"NextToken": The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

"ProvisioningStatus": The status object for the permission set provisioning operation.

Main.Sso_Admin.list_tags_for_resource — Method

list_tags_for_resource(instance_arn, resource_arn)
list_tags_for_resource(instance_arn, resource_arn, params::Dict{String,<:Any})

Lists the tags that are attached to a specified resource.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

resource_arn: The ARN of the resource with the tags to be listed.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"NextToken": The pagination token for the list API. Initially the value is null. Use the output of previous API calls to make subsequent calls.

Main.Sso_Admin.provision_permission_set — Method

provision_permission_set(instance_arn, permission_set_arn, target_type)
provision_permission_set(instance_arn, permission_set_arn, target_type, params::Dict{String,<:Any})

The process by which a specified permission set is provisioned to the specified target.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

permission_set_arn: The ARN of the permission set.

target_type: The entity type for which the assignment will be created.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"TargetId": TargetID is an AWS account identifier, typically a 10-12 digit string (For example, 123456789012).

Main.Sso_Admin.put_inline_policy_to_permission_set — Method

put_inline_policy_to_permission_set(inline_policy, instance_arn, permission_set_arn)
put_inline_policy_to_permission_set(inline_policy, instance_arn, permission_set_arn, params::Dict{String,<:Any})

Attaches an inline policy to a permission set. If the permission set is already referenced by one or more account assignments, you will need to call ProvisionPermissionSet after this action to apply the corresponding IAM policy updates to all assigned accounts.

Arguments

inline_policy: The inline policy to attach to a PermissionSet.

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

permission_set_arn: The ARN of the permission set.

Main.Sso_Admin.put_permissions_boundary_to_permission_set — Method

put_permissions_boundary_to_permission_set(instance_arn, permission_set_arn, permissions_boundary)
put_permissions_boundary_to_permission_set(instance_arn, permission_set_arn, permissions_boundary, params::Dict{String,<:Any})

Attaches an AWS managed or customer managed policy to the specified PermissionSet as a permissions boundary.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed.

permission_set_arn: The ARN of the PermissionSet.

permissions_boundary: The permissions boundary that you want to attach to a PermissionSet.

Main.Sso_Admin.tag_resource — Method

tag_resource(instance_arn, resource_arn, tags)
tag_resource(instance_arn, resource_arn, tags, params::Dict{String,<:Any})

Associates a set of tags with a specified resource.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

resource_arn: The ARN of the resource with the tags to be listed.

tags: A set of key-value pairs that are used to manage the resource.

Main.Sso_Admin.untag_resource — Method

untag_resource(instance_arn, resource_arn, tag_keys)
untag_resource(instance_arn, resource_arn, tag_keys, params::Dict{String,<:Any})

Disassociates a set of tags from a specified resource.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

resource_arn: The ARN of the resource with the tags to be listed.

tag_keys: The keys of tags that are attached to the resource.

Main.Sso_Admin.update_instance_access_control_attribute_configuration — Method

update_instance_access_control_attribute_configuration(instance_access_control_attribute_configuration, instance_arn)
update_instance_access_control_attribute_configuration(instance_access_control_attribute_configuration, instance_arn, params::Dict{String,<:Any})

Updates the IAM Identity Center identity store attributes that you can use with the IAM Identity Center instance for attributes-based access control (ABAC). When using an external identity provider as an identity source, you can pass attributes through the SAML assertion as an alternative to configuring attributes from the IAM Identity Center identity store. If a SAML assertion passes any of these attributes, IAM Identity Center replaces the attribute value with the value from the IAM Identity Center identity store. For more information about ABAC, see Attribute-Based Access Control in the IAM Identity Center User Guide.

Arguments

instance_access_control_attribute_configuration: Updates the attributes for your ABAC configuration.

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed.

Main.Sso_Admin.update_permission_set — Method

update_permission_set(instance_arn, permission_set_arn)
update_permission_set(instance_arn, permission_set_arn, params::Dict{String,<:Any})

Updates an existing permission set.

Arguments

instance_arn: The ARN of the IAM Identity Center instance under which the operation will be executed. For more information about ARNs, see Amazon Resource Names (ARNs) and AWS Service Namespaces in the AWS General Reference.

permission_set_arn: The ARN of the permission set.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": The description of the PermissionSet.

"RelayState": Used to redirect users within the application during the federation authentication process.

"SessionDuration": The length of time that the application user sessions are valid for in the ISO-8601 standard.