Supplychain

This page documents function available when using the Supplychain module, created with @service Supplychain.

Index

Main.Supplychain.create_bill_of_materials_import_job

Main.Supplychain.create_data_integration_flow

Main.Supplychain.create_data_lake_dataset

Main.Supplychain.create_data_lake_namespace

Main.Supplychain.create_instance

Main.Supplychain.delete_data_integration_flow

Main.Supplychain.delete_data_lake_dataset

Main.Supplychain.delete_data_lake_namespace

Main.Supplychain.delete_instance

Main.Supplychain.get_bill_of_materials_import_job

Main.Supplychain.get_data_integration_event

Main.Supplychain.get_data_integration_flow

Main.Supplychain.get_data_integration_flow_execution

Main.Supplychain.get_data_lake_dataset

Main.Supplychain.get_data_lake_namespace

Main.Supplychain.get_instance

Main.Supplychain.list_data_integration_events

Main.Supplychain.list_data_integration_flow_executions

Main.Supplychain.list_data_integration_flows

Main.Supplychain.list_data_lake_datasets

Main.Supplychain.list_data_lake_namespaces

Main.Supplychain.list_instances

Main.Supplychain.list_tags_for_resource

Main.Supplychain.send_data_integration_event

Main.Supplychain.tag_resource

Main.Supplychain.untag_resource

Main.Supplychain.update_data_integration_flow

Main.Supplychain.update_data_lake_dataset

Main.Supplychain.update_data_lake_namespace

Main.Supplychain.update_instance

Documentation

Main.Supplychain.create_bill_of_materials_import_job — Function

create_bill_of_materials_import_job(instance_id, s3uri)
create_bill_of_materials_import_job(instance_id, s3uri, params::Dict{String,<:Any})

CreateBillOfMaterialsImportJob creates an import job for the Product Bill Of Materials (BOM) entity. For information on the product_bom entity, see the AWS Supply Chain User Guide.

The CSV file must be located in an Amazon S3 location accessible to AWS Supply Chain. It is recommended to use the same Amazon S3 bucket created during your AWS Supply Chain instance creation.

Arguments

instance_id: The AWS Supply Chain instance identifier.

s3uri: The S3 URI of the CSV file to be imported. The bucket must grant permissions for AWS Supply Chain to read the file.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": An idempotency token ensures the API request is only completed no more than once. This way, retrying the request will not trigger the operation multiple times. A client token is a unique, case-sensitive string of 33 to 128 ASCII characters. To make an idempotent API request, specify a client token in the request. You should not reuse the same client token for other requests. If you retry a successful request with the same client token, the request will succeed with no further actions being taken, and you will receive the same API response as the original successful request.

Main.Supplychain.create_data_integration_flow — Function

create_data_integration_flow(instance_id, name, sources, target, transformation)
create_data_integration_flow(instance_id, name, sources, target, transformation, params::Dict{String,<:Any})

Enables you to programmatically create a data pipeline to ingest data from source systems such as Amazon S3 buckets, to a predefined Amazon Web Services Supply Chain dataset (product, inbound_order) or a temporary dataset along with the data transformation query provided with the API.

Arguments

instance_id: The Amazon Web Services Supply Chain instance identifier.

name: Name of the DataIntegrationFlow.

sources: The source configurations for DataIntegrationFlow.

target: The target configurations for DataIntegrationFlow.

transformation: The transformation configurations for DataIntegrationFlow.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"tags": The tags of the DataIntegrationFlow to be created

Main.Supplychain.create_data_lake_dataset — Function

create_data_lake_dataset(instance_id, name, namespace)
create_data_lake_dataset(instance_id, name, namespace, params::Dict{String,<:Any})

Enables you to programmatically create an Amazon Web Services Supply Chain data lake dataset. Developers can create the datasets using their pre-defined or custom schema for a given instance ID, namespace, and dataset name.

Arguments

instance_id: The Amazon Web Services Supply Chain instance identifier.

name: The name of the dataset. For asc name space, the name must be one of the supported data entities under https://docs.aws.amazon.com/aws-supply-chain/latest/userguide/data-model-asc.html.

namespace: The namespace of the dataset, besides the custom defined namespace, every instance comes with below pre-defined namespaces:

asc - For information on the Amazon Web Services Supply Chain supported datasets see https://docs.aws.amazon.com/aws-supply-chain/latest/userguide/data-model-asc.html.

default - For datasets with custom user-defined schemas.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The description of the dataset.

"partitionSpec": The partition specification of the dataset. Partitioning can effectively improve the dataset query performance by reducing the amount of data scanned during query execution. But partitioning or not will affect how data get ingested by data ingestion methods, such as SendDataIntegrationEvent's dataset UPSERT will upsert records within partition (instead of within whole dataset). For more details, refer to those data ingestion documentations.

"schema": The custom schema of the data lake dataset and required for dataset in default and custom namespaces.

"tags": The tags of the dataset.

Main.Supplychain.create_data_lake_namespace — Function

create_data_lake_namespace(instance_id, name)
create_data_lake_namespace(instance_id, name, params::Dict{String,<:Any})

Enables you to programmatically create an Amazon Web Services Supply Chain data lake namespace. Developers can create the namespaces for a given instance ID.

Arguments

instance_id: The Amazon Web Services Supply Chain instance identifier.

name: The name of the namespace. Noted you cannot create namespace with name starting with asc, default, scn, aws, amazon, amzn

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The description of the namespace.

"tags": The tags of the namespace.

Main.Supplychain.create_instance — Function

create_instance()
create_instance(params::Dict{String,<:Any})

Enables you to programmatically create an Amazon Web Services Supply Chain instance by applying KMS keys and relevant information associated with the API without using the Amazon Web Services console.

This is an asynchronous operation. Upon receiving a CreateInstance request, Amazon Web Services Supply Chain immediately returns the instance resource, instance ID, and the initializing state while simultaneously creating all required Amazon Web Services resources for an instance creation. You can use GetInstance to check the status of the instance. If the instance results in an unhealthy state, you need to check the error message, delete the current instance, and recreate a new one based on the mitigation from the error message.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": The client token for idempotency.

"instanceDescription": The AWS Supply Chain instance description.

"instanceName": The AWS Supply Chain instance name.

"kmsKeyArn": The ARN (Amazon Resource Name) of the Key Management Service (KMS) key you provide for encryption. This is required if you do not want to use the Amazon Web Services owned KMS key. If you don't provide anything here, AWS Supply Chain uses the Amazon Web Services owned KMS key.

"tags": The Amazon Web Services tags of an instance to be created.

"webAppDnsDomain": The DNS subdomain of the web app. This would be "example" in the URL "example.scn.global.on.aws". You can set this to a custom value, as long as the domain isn't already being used by someone else. The name may only include alphanumeric characters and hyphens.

Main.Supplychain.delete_data_integration_flow — Function

delete_data_integration_flow(instance_id, name)
delete_data_integration_flow(instance_id, name, params::Dict{String,<:Any})

Enable you to programmatically delete an existing data pipeline for the provided Amazon Web Services Supply Chain instance and DataIntegrationFlow name.

Arguments

instance_id: The Amazon Web Services Supply Chain instance identifier.

name: The name of the DataIntegrationFlow to be deleted.

Main.Supplychain.delete_data_lake_dataset — Function

delete_data_lake_dataset(instance_id, name, namespace)
delete_data_lake_dataset(instance_id, name, namespace, params::Dict{String,<:Any})

Enables you to programmatically delete an Amazon Web Services Supply Chain data lake dataset. Developers can delete the existing datasets for a given instance ID, namespace, and instance name.

Arguments

instance_id: The AWS Supply Chain instance identifier.

name: The name of the dataset. For asc namespace, the name must be one of the supported data entities under https://docs.aws.amazon.com/aws-supply-chain/latest/userguide/data-model-asc.html.

namespace: The namespace of the dataset, besides the custom defined namespace, every instance comes with below pre-defined namespaces:

asc - For information on the Amazon Web Services Supply Chain supported datasets see https://docs.aws.amazon.com/aws-supply-chain/latest/userguide/data-model-asc.html.

default - For datasets with custom user-defined schemas.

Main.Supplychain.delete_data_lake_namespace — Function

delete_data_lake_namespace(instance_id, name)
delete_data_lake_namespace(instance_id, name, params::Dict{String,<:Any})

Enables you to programmatically delete an Amazon Web Services Supply Chain data lake namespace and its underling datasets. Developers can delete the existing namespaces for a given instance ID and namespace name.

Arguments

instance_id: The AWS Supply Chain instance identifier.

name: The name of the namespace. Noted you cannot delete pre-defined namespace like asc, default which are only deleted through instance deletion.

Main.Supplychain.delete_instance — Function

delete_instance(instance_id)
delete_instance(instance_id, params::Dict{String,<:Any})

Enables you to programmatically delete an Amazon Web Services Supply Chain instance by deleting the KMS keys and relevant information associated with the API without using the Amazon Web Services console.

This is an asynchronous operation. Upon receiving a DeleteInstance request, Amazon Web Services Supply Chain immediately returns a response with the instance resource, delete state while cleaning up all Amazon Web Services resources created during the instance creation process. You can use the GetInstance action to check the instance status.

Arguments

instance_id: The AWS Supply Chain instance identifier.

Main.Supplychain.get_bill_of_materials_import_job — Function

get_bill_of_materials_import_job(instance_id, job_id)
get_bill_of_materials_import_job(instance_id, job_id, params::Dict{String,<:Any})

Get status and details of a BillOfMaterialsImportJob.

Arguments

instance_id: The AWS Supply Chain instance identifier.

job_id: The BillOfMaterialsImportJob identifier.

Main.Supplychain.get_data_integration_event — Function

get_data_integration_event(event_id, instance_id)
get_data_integration_event(event_id, instance_id, params::Dict{String,<:Any})

Enables you to programmatically view an Amazon Web Services Supply Chain Data Integration Event. Developers can view the eventType, eventGroupId, eventTimestamp, datasetTarget, datasetLoadExecution.

Arguments

event_id: The unique event identifier.

instance_id: The Amazon Web Services Supply Chain instance identifier.

Main.Supplychain.get_data_integration_flow — Function

get_data_integration_flow(instance_id, name)
get_data_integration_flow(instance_id, name, params::Dict{String,<:Any})

Enables you to programmatically view a specific data pipeline for the provided Amazon Web Services Supply Chain instance and DataIntegrationFlow name.

Arguments

instance_id: The Amazon Web Services Supply Chain instance identifier.

name: The name of the DataIntegrationFlow created.

Main.Supplychain.get_data_integration_flow_execution — Function

get_data_integration_flow_execution(execution_id, flow_name, instance_id)
get_data_integration_flow_execution(execution_id, flow_name, instance_id, params::Dict{String,<:Any})

Get the flow execution.

Arguments

execution_id: The flow execution identifier.

flow_name: The flow name.

instance_id: The AWS Supply Chain instance identifier.

Main.Supplychain.get_data_lake_dataset — Function

get_data_lake_dataset(instance_id, name, namespace)
get_data_lake_dataset(instance_id, name, namespace, params::Dict{String,<:Any})

Enables you to programmatically view an Amazon Web Services Supply Chain data lake dataset. Developers can view the data lake dataset information such as namespace, schema, and so on for a given instance ID, namespace, and dataset name.

Arguments

instance_id: The Amazon Web Services Supply Chain instance identifier.

name: The name of the dataset. For asc namespace, the name must be one of the supported data entities under https://docs.aws.amazon.com/aws-supply-chain/latest/userguide/data-model-asc.html.

namespace: The namespace of the dataset, besides the custom defined namespace, every instance comes with below pre-defined namespaces:

asc - For information on the Amazon Web Services Supply Chain supported datasets see https://docs.aws.amazon.com/aws-supply-chain/latest/userguide/data-model-asc.html.

default - For datasets with custom user-defined schemas.

Main.Supplychain.get_data_lake_namespace — Function

get_data_lake_namespace(instance_id, name)
get_data_lake_namespace(instance_id, name, params::Dict{String,<:Any})

Enables you to programmatically view an Amazon Web Services Supply Chain data lake namespace. Developers can view the data lake namespace information such as description for a given instance ID and namespace name.

Arguments

instance_id: The Amazon Web Services Supply Chain instance identifier.

name: The name of the namespace. Besides the namespaces user created, you can also specify the pre-defined namespaces:

asc - Pre-defined namespace containing Amazon Web Services Supply Chain supported datasets, see https://docs.aws.amazon.com/aws-supply-chain/latest/userguide/data-model-asc.html.

default - Pre-defined namespace containing datasets with custom user-defined schemas.

Main.Supplychain.get_instance — Function

get_instance(instance_id)
get_instance(instance_id, params::Dict{String,<:Any})

Enables you to programmatically retrieve the information related to an Amazon Web Services Supply Chain instance ID.

Arguments

instance_id: The AWS Supply Chain instance identifier

Main.Supplychain.list_data_integration_events — Function

list_data_integration_events(instance_id)
list_data_integration_events(instance_id, params::Dict{String,<:Any})

Enables you to programmatically list all data integration events for the provided Amazon Web Services Supply Chain instance.

Arguments

instance_id: The Amazon Web Services Supply Chain instance identifier.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"eventType": List data integration events for the specified eventType.

"maxResults": Specify the maximum number of data integration events to fetch in one paginated request.

"nextToken": The pagination token to fetch the next page of the data integration events.

Main.Supplychain.list_data_integration_flow_executions — Function

list_data_integration_flow_executions(flow_name, instance_id)
list_data_integration_flow_executions(flow_name, instance_id, params::Dict{String,<:Any})

List flow executions.

Arguments

flow_name: The flow name.

instance_id: The AWS Supply Chain instance identifier.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The number to specify the max number of flow executions to fetch in this paginated request.

"nextToken": The pagination token to fetch next page of flow executions.

Main.Supplychain.list_data_integration_flows — Function

list_data_integration_flows(instance_id)
list_data_integration_flows(instance_id, params::Dict{String,<:Any})

Enables you to programmatically list all data pipelines for the provided Amazon Web Services Supply Chain instance.

Arguments

instance_id: The Amazon Web Services Supply Chain instance identifier.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Specify the maximum number of DataIntegrationFlows to fetch in one paginated request.

"nextToken": The pagination token to fetch the next page of the DataIntegrationFlows.

Main.Supplychain.list_data_lake_datasets — Function

list_data_lake_datasets(instance_id, namespace)
list_data_lake_datasets(instance_id, namespace, params::Dict{String,<:Any})

Enables you to programmatically view the list of Amazon Web Services Supply Chain data lake datasets. Developers can view the datasets and the corresponding information such as namespace, schema, and so on for a given instance ID and namespace.

Arguments

instance_id: The Amazon Web Services Supply Chain instance identifier.

namespace: The namespace of the dataset, besides the custom defined namespace, every instance comes with below pre-defined namespaces:

asc - For information on the Amazon Web Services Supply Chain supported datasets see https://docs.aws.amazon.com/aws-supply-chain/latest/userguide/data-model-asc.html.

default - For datasets with custom user-defined schemas.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The max number of datasets to fetch in this paginated request.

"nextToken": The pagination token to fetch next page of datasets.

Main.Supplychain.list_data_lake_namespaces — Function

list_data_lake_namespaces(instance_id)
list_data_lake_namespaces(instance_id, params::Dict{String,<:Any})

Enables you to programmatically view the list of Amazon Web Services Supply Chain data lake namespaces. Developers can view the namespaces and the corresponding information such as description for a given instance ID. Note that this API only return custom namespaces, instance pre-defined namespaces are not included.

Arguments

instance_id: The Amazon Web Services Supply Chain instance identifier.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The max number of namespaces to fetch in this paginated request.

"nextToken": The pagination token to fetch next page of namespaces.

Main.Supplychain.list_instances — Function

list_instances()
list_instances(params::Dict{String,<:Any})

List all Amazon Web Services Supply Chain instances for a specific account. Enables you to programmatically list all Amazon Web Services Supply Chain instances based on their account ID, instance name, and state of the instance (active or delete).

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"instanceNameFilter": The filter to ListInstances based on their names.

"instanceStateFilter": The filter to ListInstances based on their state.

"maxResults": Specify the maximum number of instances to fetch in this paginated request.

"nextToken": The pagination token to fetch the next page of instances.

Main.Supplychain.list_tags_for_resource — Function

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

List all the tags for an Amazon Web ServicesSupply Chain resource. You can list all the tags added to a resource. By listing the tags, developers can view the tag level information on a resource and perform actions such as, deleting a resource associated with a particular tag.

Arguments

resource_arn: The Amazon Web Services Supply chain resource ARN that needs tags to be listed.

Main.Supplychain.send_data_integration_event — Function

send_data_integration_event(data, event_group_id, event_type, instance_id)
send_data_integration_event(data, event_group_id, event_type, instance_id, params::Dict{String,<:Any})

Send the data payload for the event with real-time data for analysis or monitoring. The real-time data events are stored in an Amazon Web Services service before being processed and stored in data lake.

Arguments

data: The data payload of the event, should follow the data schema of the target dataset, or see Data entities supported in AWS Supply Chain. To send single data record, use JsonObject format; to send multiple data records, use JsonArray format.

Note that for AWS Supply Chain dataset under asc namespace, it has a connection_id internal field that is not allowed to be provided by client directly, they will be auto populated.

event_group_id: Event identifier (for example, orderId for InboundOrder) used for data sharding or partitioning. Noted under one eventGroupId of same eventType and instanceId, events are processed sequentially in the order they are received by the server.

event_type: The data event type.

scn.data.dataset - Send data directly to any specified dataset.

scn.data.supplyplan - Send data to supply_plan dataset.

scn.data.shipmentstoporder - Send data to shipmentstoporder dataset.

scn.data.shipmentstop - Send data to shipment_stop dataset.

scn.data.shipment - Send data to shipment dataset.

scn.data.reservation - Send data to reservation dataset.

scn.data.processproduct - Send data to process_product dataset.

scn.data.processoperation - Send data to process_operation dataset.

scn.data.processheader - Send data to process_header dataset.

scn.data.forecast - Send data to forecast dataset.

scn.data.inventorylevel - Send data to inv_level dataset.

scn.data.inboundorder - Send data to inbound_order dataset.

scn.data.inboundorderline - Send data to inboundorderline dataset.

scn.data.inboundorderlineschedule - Send data to inboundorderline_schedule dataset.

scn.data.outboundorderline - Send data to outboundorderline dataset.

scn.data.outboundshipment - Send data to outbound_shipment dataset.

instance_id: The AWS Supply Chain instance identifier.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": The idempotent client token. The token is active for 8 hours, and within its lifetime, it ensures the request completes only once upon retry with same client token. If omitted, the AWS SDK generates a unique value so that AWS SDK can safely retry the request upon network errors.

"datasetTarget": The target dataset configuration for scn.data.dataset event type.

"eventTimestamp": The timestamp (in epoch seconds) associated with the event. If not provided, it will be assigned with current timestamp.

Main.Supplychain.tag_resource — Function

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

You can create tags during or after creating a resource such as instance, data flow, or dataset in AWS Supply chain. During the data ingestion process, you can add tags such as dev, test, or prod to data flows created during the data ingestion process in the AWS Supply Chain datasets. You can use these tags to identify a group of resources or a single resource used by the developer.

Arguments

resource_arn: The Amazon Web Services Supply chain resource ARN that needs to be tagged.

tags: The tags of the Amazon Web Services Supply chain resource to be created.

Main.Supplychain.untag_resource — Function

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

You can delete tags for an Amazon Web Services Supply chain resource such as instance, data flow, or dataset in AWS Supply Chain. During the data ingestion process, you can delete tags such as dev, test, or prod to data flows created during the data ingestion process in the AWS Supply Chain datasets.

Arguments

resource_arn: The Amazon Web Services Supply chain resource ARN that needs to be untagged.

tag_keys: The list of tag keys to be deleted for an Amazon Web Services Supply Chain resource.

Main.Supplychain.update_data_integration_flow — Function

update_data_integration_flow(instance_id, name)
update_data_integration_flow(instance_id, name, params::Dict{String,<:Any})

Enables you to programmatically update an existing data pipeline to ingest data from the source systems such as, Amazon S3 buckets, to a predefined Amazon Web Services Supply Chain dataset (product, inbound_order) or a temporary dataset along with the data transformation query provided with the API.

Arguments

instance_id: The Amazon Web Services Supply Chain instance identifier.

name: The name of the DataIntegrationFlow to be updated.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"sources": The new source configurations for the DataIntegrationFlow.

"target": The new target configurations for the DataIntegrationFlow.

"transformation": The new transformation configurations for the DataIntegrationFlow.

Main.Supplychain.update_data_lake_dataset — Function

update_data_lake_dataset(instance_id, name, namespace)
update_data_lake_dataset(instance_id, name, namespace, params::Dict{String,<:Any})

Enables you to programmatically update an Amazon Web Services Supply Chain data lake dataset. Developers can update the description of a data lake dataset for a given instance ID, namespace, and dataset name.

Arguments

instance_id: The Amazon Web Services Chain instance identifier.

name: The name of the dataset. For asc namespace, the name must be one of the supported data entities under https://docs.aws.amazon.com/aws-supply-chain/latest/userguide/data-model-asc.html.

namespace: The namespace of the dataset, besides the custom defined namespace, every instance comes with below pre-defined namespaces:

asc - For information on the Amazon Web Services Supply Chain supported datasets see https://docs.aws.amazon.com/aws-supply-chain/latest/userguide/data-model-asc.html.

default - For datasets with custom user-defined schemas.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The updated description of the data lake dataset.

Main.Supplychain.update_data_lake_namespace — Function

update_data_lake_namespace(instance_id, name)
update_data_lake_namespace(instance_id, name, params::Dict{String,<:Any})

Enables you to programmatically update an Amazon Web Services Supply Chain data lake namespace. Developers can update the description of a data lake namespace for a given instance ID and namespace name.

Arguments

instance_id: The Amazon Web Services Chain instance identifier.

name: The name of the namespace. Noted you cannot update namespace with name starting with asc, default, scn, aws, amazon, amzn

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The updated description of the data lake namespace.

Main.Supplychain.update_instance — Function

update_instance(instance_id)
update_instance(instance_id, params::Dict{String,<:Any})

Enables you to programmatically update an Amazon Web Services Supply Chain instance description by providing all the relevant information such as account ID, instance ID and so on without using the AWS console.

Arguments

instance_id: The AWS Supply Chain instance identifier.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"instanceDescription": The AWS Supply Chain instance description.

"instanceName": The AWS Supply Chain instance name.