Timestream Influxdb

This page documents function available when using the Timestream_Influxdb module, created with @service Timestream_Influxdb.

Index

Main.Timestream_Influxdb.create_db_instance

Main.Timestream_Influxdb.create_db_parameter_group

Main.Timestream_Influxdb.delete_db_instance

Main.Timestream_Influxdb.get_db_instance

Main.Timestream_Influxdb.get_db_parameter_group

Main.Timestream_Influxdb.list_db_instances

Main.Timestream_Influxdb.list_db_parameter_groups

Main.Timestream_Influxdb.list_tags_for_resource

Main.Timestream_Influxdb.tag_resource

Main.Timestream_Influxdb.untag_resource

Main.Timestream_Influxdb.update_db_instance

Documentation

Main.Timestream_Influxdb.create_db_instance — Method

create_db_instance(allocated_storage, db_instance_type, name, password, vpc_security_group_ids, vpc_subnet_ids)
create_db_instance(allocated_storage, db_instance_type, name, password, vpc_security_group_ids, vpc_subnet_ids, params::Dict{String,<:Any})

Creates a new Timestream for InfluxDB DB instance.

Arguments

allocated_storage: The amount of storage to allocate for your DB storage type in GiB (gibibytes).

db_instance_type: The Timestream for InfluxDB DB instance type to run InfluxDB on.

name: The name that uniquely identifies the DB instance when interacting with the Amazon Timestream for InfluxDB API and CLI commands. This name will also be a prefix included in the endpoint. DB instance names must be unique per customer and per region.

password: The password of the initial admin user created in InfluxDB. This password will allow you to access the InfluxDB UI to perform various administrative tasks and also use the InfluxDB CLI to create an operator token. These attributes will be stored in a Secret created in AWS SecretManager in your account.

vpc_security_group_ids: A list of VPC security group IDs to associate with the DB instance.

vpc_subnet_ids: A list of VPC subnet IDs to associate with the DB instance. Provide at least two VPC subnet IDs in different availability zones when deploying with a Multi-AZ standby.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"bucket": The name of the initial InfluxDB bucket. All InfluxDB data is stored in a bucket. A bucket combines the concept of a database and a retention period (the duration of time that each data point persists). A bucket belongs to an organization.

"dbParameterGroupIdentifier": The id of the DB parameter group to assign to your DB instance. DB parameter groups specify how the database is configured. For example, DB parameter groups can specify the limit for query concurrency.

"dbStorageType": The Timestream for InfluxDB DB storage type to read and write InfluxDB data. You can choose between 3 different types of provisioned Influx IOPS included storage according to your workloads requirements: Influx IO Included 3000 IOPS Influx IO Included 12000 IOPS Influx IO Included 16000 IOPS

"deploymentType": Specifies whether the DB instance will be deployed as a standalone instance or with a Multi-AZ standby for high availability.

"logDeliveryConfiguration": Configuration for sending InfluxDB engine logs to a specified S3 bucket.

"organization": The name of the initial organization for the initial admin user in InfluxDB. An InfluxDB organization is a workspace for a group of users.

"publiclyAccessible": Configures the DB instance with a public IP to facilitate access.

"tags": A list of key-value pairs to associate with the DB instance.

"username": The username of the initial admin user created in InfluxDB. Must start with a letter and can't end with a hyphen or contain two consecutive hyphens. For example, my-user1. This username will allow you to access the InfluxDB UI to perform various administrative tasks and also use the InfluxDB CLI to create an operator token. These attributes will be stored in a Secret created in Amazon Secrets Manager in your account.

Main.Timestream_Influxdb.create_db_parameter_group — Method

create_db_parameter_group(name)
create_db_parameter_group(name, params::Dict{String,<:Any})

Creates a new Timestream for InfluxDB DB parameter group to associate with DB instances.

Arguments

name: The name of the DB parameter group. The name must be unique per customer and per region.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A description of the DB parameter group.

"parameters": A list of the parameters that comprise the DB parameter group.

"tags": A list of key-value pairs to associate with the DB parameter group.

Main.Timestream_Influxdb.delete_db_instance — Method

delete_db_instance(identifier)
delete_db_instance(identifier, params::Dict{String,<:Any})

Deletes a Timestream for InfluxDB DB instance.

Arguments

identifier: The id of the DB instance.

Main.Timestream_Influxdb.get_db_instance — Method

get_db_instance(identifier)
get_db_instance(identifier, params::Dict{String,<:Any})

Returns a Timestream for InfluxDB DB instance.

Arguments

identifier: The id of the DB instance.

Main.Timestream_Influxdb.get_db_parameter_group — Method

get_db_parameter_group(identifier)
get_db_parameter_group(identifier, params::Dict{String,<:Any})

Returns a Timestream for InfluxDB DB parameter group.

Arguments

identifier: The id of the DB parameter group.

Main.Timestream_Influxdb.list_db_instances — Method

list_db_instances()
list_db_instances(params::Dict{String,<:Any})

Returns a list of Timestream for InfluxDB DB instances.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of items to return in the output. If the total number of items available is more than the value specified, a NextToken is provided in the output. To resume pagination, provide the NextToken value as argument of a subsequent API invocation.

"nextToken": The pagination token. To resume pagination, provide the NextToken value as argument of a subsequent API invocation.

Main.Timestream_Influxdb.list_db_parameter_groups — Method

list_db_parameter_groups()
list_db_parameter_groups(params::Dict{String,<:Any})

Returns a list of Timestream for InfluxDB DB parameter groups.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of items to return in the output. If the total number of items available is more than the value specified, a NextToken is provided in the output. To resume pagination, provide the NextToken value as argument of a subsequent API invocation.

"nextToken": The pagination token. To resume pagination, provide the NextToken value as argument of a subsequent API invocation.

Main.Timestream_Influxdb.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

A list of tags applied to the resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the tagged resource.

Main.Timestream_Influxdb.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Tags are composed of a Key/Value pairs. You can use tags to categorize and track your Timestream for InfluxDB resources.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the tagged resource.

tags: A list of tags used to categorize and track resources.

Main.Timestream_Influxdb.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes the tag from the specified resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the tagged resource.

tag_keys: The keys used to identify the tags.

Main.Timestream_Influxdb.update_db_instance — Method

update_db_instance(identifier)
update_db_instance(identifier, params::Dict{String,<:Any})

Updates a Timestream for InfluxDB DB instance.

Arguments

identifier: The id of the DB instance.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"dbParameterGroupIdentifier": The id of the DB parameter group to assign to your DB instance. DB parameter groups specify how the database is configured. For example, DB parameter groups can specify the limit for query concurrency.

"logDeliveryConfiguration": Configuration for sending InfluxDB engine logs to send to specified S3 bucket.