Timestream Write

This page documents function available when using the Timestream_Write module, created with @service Timestream_Write.

Index

Main.Timestream_Write.create_database

Main.Timestream_Write.create_table

Main.Timestream_Write.delete_database

Main.Timestream_Write.delete_table

Main.Timestream_Write.describe_database

Main.Timestream_Write.describe_endpoints

Main.Timestream_Write.describe_table

Main.Timestream_Write.list_databases

Main.Timestream_Write.list_tables

Main.Timestream_Write.list_tags_for_resource

Main.Timestream_Write.tag_resource

Main.Timestream_Write.untag_resource

Main.Timestream_Write.update_database

Main.Timestream_Write.update_table

Main.Timestream_Write.write_records

Documentation

Main.Timestream_Write.create_database — Method

create_database(database_name)
create_database(database_name, params::Dict{String,<:Any})

Creates a new Timestream database. If the KMS key is not specified, the database will be encrypted with a Timestream managed KMS key located in your account. Refer to Amazon Web Services managed KMS keys for more info. Service quotas apply. See code sample for details.

Arguments

database_name: The name of the Timestream database.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"KmsKeyId": The KMS key for the database. If the KMS key is not specified, the database will be encrypted with a Timestream managed KMS key located in your account. Refer to Amazon Web Services managed KMS keys for more info.

"Tags": A list of key-value pairs to label the table.

Main.Timestream_Write.create_table — Method

create_table(database_name, table_name)
create_table(database_name, table_name, params::Dict{String,<:Any})

The CreateTable operation adds a new table to an existing database in your account. In an Amazon Web Services account, table names must be at least unique within each Region if they are in the same database. You may have identical table names in the same Region if the tables are in separate databases. While creating the table, you must specify the table name, database name, and the retention properties. Service quotas apply. See code sample for details.

Arguments

database_name: The name of the Timestream database.

table_name: The name of the Timestream table.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MagneticStoreWriteProperties": Contains properties to set on the table when enabling magnetic store writes.

"RetentionProperties": The duration for which your time series data must be stored in the memory store and the magnetic store.

"Tags": A list of key-value pairs to label the table.

Main.Timestream_Write.delete_database — Method

delete_database(database_name)
delete_database(database_name, params::Dict{String,<:Any})

Deletes a given Timestream database. This is an irreversible operation. After a database is deleted, the time series data from its tables cannot be recovered. All tables in the database must be deleted first, or a ValidationException error will be thrown. Due to the nature of distributed retries, the operation can return either success or a ResourceNotFoundException. Clients should consider them equivalent. See code sample for details.

Arguments

database_name: The name of the Timestream database to be deleted.

Main.Timestream_Write.delete_table — Method

delete_table(database_name, table_name)
delete_table(database_name, table_name, params::Dict{String,<:Any})

Deletes a given Timestream table. This is an irreversible operation. After a Timestream database table is deleted, the time series data stored in the table cannot be recovered. Due to the nature of distributed retries, the operation can return either success or a ResourceNotFoundException. Clients should consider them equivalent. See code sample for details.

Arguments

database_name: The name of the database where the Timestream database is to be deleted.

table_name: The name of the Timestream table to be deleted.

Main.Timestream_Write.describe_database — Method

describe_database(database_name)
describe_database(database_name, params::Dict{String,<:Any})

Returns information about the database, including the database name, time that the database was created, and the total number of tables found within the database. Service quotas apply. See code sample for details.

Arguments

database_name: The name of the Timestream database.

Main.Timestream_Write.describe_endpoints — Method

describe_endpoints()
describe_endpoints(params::Dict{String,<:Any})

DescribeEndpoints returns a list of available endpoints to make Timestream API calls against. This API is available through both Write and Query. Because the Timestream SDKs are designed to transparently work with the service’s architecture, including the management and mapping of the service endpoints, it is not recommended that you use this API unless: You are using VPC endpoints (Amazon Web Services PrivateLink) with Timestream Your application uses a programming language that does not yet have SDK support You require better control over the client-side implementation For detailed information on how and when to use and implement DescribeEndpoints, see The Endpoint Discovery Pattern.

Main.Timestream_Write.describe_table — Method

describe_table(database_name, table_name)
describe_table(database_name, table_name, params::Dict{String,<:Any})

Returns information about the table, including the table name, database name, retention duration of the memory store and the magnetic store. Service quotas apply. See code sample for details.

Arguments

database_name: The name of the Timestream database.

table_name: The name of the Timestream table.

Main.Timestream_Write.list_databases — Method

list_databases()
list_databases(params::Dict{String,<:Any})

Returns a list of your Timestream databases. Service quotas apply. See code sample for details.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The total number of items to return in the output. If the total number of items available is more than the value specified, a NextToken is provided in the output. To resume pagination, provide the NextToken value as argument of a subsequent API invocation.

"NextToken": The pagination token. To resume pagination, provide the NextToken value as argument of a subsequent API invocation.

Main.Timestream_Write.list_tables — Method

list_tables()
list_tables(params::Dict{String,<:Any})

A list of tables, along with the name, status and retention properties of each table. See code sample for details.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DatabaseName": The name of the Timestream database.

"MaxResults": The total number of items to return in the output. If the total number of items available is more than the value specified, a NextToken is provided in the output. To resume pagination, provide the NextToken value as argument of a subsequent API invocation.

"NextToken": The pagination token. To resume pagination, provide the NextToken value as argument of a subsequent API invocation.

Main.Timestream_Write.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

List all tags on a Timestream resource.

Arguments

resource_arn: The Timestream resource with tags to be listed. This value is an Amazon Resource Name (ARN).

Main.Timestream_Write.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Associate a set of tags with a Timestream resource. You can then activate these user-defined tags so that they appear on the Billing and Cost Management console for cost allocation tracking.

Arguments

resource_arn: Identifies the Timestream resource to which tags should be added. This value is an Amazon Resource Name (ARN).

tags: The tags to be assigned to the Timestream resource.

Main.Timestream_Write.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes the association of tags from a Timestream resource.

Arguments

resource_arn: The Timestream resource that the tags will be removed from. This value is an Amazon Resource Name (ARN).

tag_keys: A list of tags keys. Existing tags of the resource whose keys are members of this list will be removed from the Timestream resource.

Main.Timestream_Write.update_database — Method

update_database(database_name, kms_key_id)
update_database(database_name, kms_key_id, params::Dict{String,<:Any})

Modifies the KMS key for an existing database. While updating the database, you must specify the database name and the identifier of the new KMS key to be used (KmsKeyId). If there are any concurrent UpdateDatabase requests, first writer wins. See code sample for details.

Arguments

database_name: The name of the database.

kms_key_id: The identifier of the new KMS key (KmsKeyId) to be used to encrypt the data stored in the database. If the KmsKeyId currently registered with the database is the same as the KmsKeyId in the request, there will not be any update. You can specify the KmsKeyId using any of the following: Key ID: 1234abcd-12ab-34cd-56ef-1234567890ab Key ARN: arn:aws:kms:us-east-1:111122223333:key/1234abcd-12ab-34cd-56ef-1234567890ab Alias name: alias/ExampleAlias Alias ARN: arn:aws:kms:us-east-1:111122223333:alias/ExampleAlias

Main.Timestream_Write.update_table — Method

update_table(database_name, table_name)
update_table(database_name, table_name, params::Dict{String,<:Any})

Modifies the retention duration of the memory store and magnetic store for your Timestream table. Note that the change in retention duration takes effect immediately. For example, if the retention period of the memory store was initially set to 2 hours and then changed to 24 hours, the memory store will be capable of holding 24 hours of data, but will be populated with 24 hours of data 22 hours after this change was made. Timestream does not retrieve data from the magnetic store to populate the memory store. See code sample for details.

Arguments

database_name: The name of the Timestream database.

table_name: The name of the Timestream table.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MagneticStoreWriteProperties": Contains properties to set on the table when enabling magnetic store writes.

"RetentionProperties": The retention duration of the memory store and the magnetic store.

Main.Timestream_Write.write_records — Method

write_records(database_name, records, table_name)
write_records(database_name, records, table_name, params::Dict{String,<:Any})

The WriteRecords operation enables you to write your time series data into Timestream. You can specify a single data point or a batch of data points to be inserted into the system. Timestream offers you with a flexible schema that auto detects the column names and data types for your Timestream tables based on the dimension names and data types of the data points you specify when invoking writes into the database. Timestream support eventual consistency read semantics. This means that when you query data immediately after writing a batch of data into Timestream, the query results might not reflect the results of a recently completed write operation. The results may also include some stale data. If you repeat the query request after a short time, the results should return the latest data. Service quotas apply. See code sample for details. Upserts You can use the Version parameter in a WriteRecords request to update data points. Timestream tracks a version number with each record. Version defaults to 1 when not specified for the record in the request. Timestream will update an existing record’s measure value along with its Version upon receiving a write request with a higher Version number for that record. Upon receiving an update request where the measure value is the same as that of the existing record, Timestream still updates Version, if it is greater than the existing value of Version. You can update a data point as many times as desired, as long as the value of Version continuously increases. For example, suppose you write a new record without indicating Version in the request. Timestream will store this record, and set Version to 1. Now, suppose you try to update this record with a WriteRecords request of the same record with a different measure value but, like before, do not provide Version. In this case, Timestream will reject this update with a RejectedRecordsException since the updated record’s version is not greater than the existing value of Version. However, if you were to resend the update request with Version set to 2, Timestream would then succeed in updating the record’s value, and the Version would be set to 2. Next, suppose you sent a WriteRecords request with this same record and an identical measure value, but with Version set to 3. In this case, Timestream would only update Version to 3. Any further updates would need to send a version number greater than 3, or the update requests would receive a RejectedRecordsException.

Arguments

database_name: The name of the Timestream database.

records: An array of records containing the unique measure, dimension, time, and version attributes for each time series data point.

table_name: The name of the Timestream table.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"CommonAttributes": A record containing the common measure, dimension, time, and version attributes shared across all the records in the request. The measure and dimension attributes specified will be merged with the measure and dimension attributes in the records object when the data is written into Timestream. Dimensions may not overlap, or a ValidationException will be thrown. In other words, a record must contain dimensions with unique names.