Tnb

This page documents function available when using the Tnb module, created with @service Tnb.

Index

Main.Tnb.cancel_sol_network_operation

Main.Tnb.create_sol_function_package

Main.Tnb.create_sol_network_instance

Main.Tnb.create_sol_network_package

Main.Tnb.delete_sol_function_package

Main.Tnb.delete_sol_network_instance

Main.Tnb.delete_sol_network_package

Main.Tnb.get_sol_function_instance

Main.Tnb.get_sol_function_package

Main.Tnb.get_sol_function_package_content

Main.Tnb.get_sol_function_package_descriptor

Main.Tnb.get_sol_network_instance

Main.Tnb.get_sol_network_operation

Main.Tnb.get_sol_network_package

Main.Tnb.get_sol_network_package_content

Main.Tnb.get_sol_network_package_descriptor

Main.Tnb.instantiate_sol_network_instance

Main.Tnb.list_sol_function_instances

Main.Tnb.list_sol_function_packages

Main.Tnb.list_sol_network_instances

Main.Tnb.list_sol_network_operations

Main.Tnb.list_sol_network_packages

Main.Tnb.list_tags_for_resource

Main.Tnb.put_sol_function_package_content

Main.Tnb.put_sol_network_package_content

Main.Tnb.tag_resource

Main.Tnb.terminate_sol_network_instance

Main.Tnb.untag_resource

Main.Tnb.update_sol_function_package

Main.Tnb.update_sol_network_instance

Main.Tnb.update_sol_network_package

Main.Tnb.validate_sol_function_package_content

Main.Tnb.validate_sol_network_package_content

Documentation

Main.Tnb.cancel_sol_network_operation — Method

cancel_sol_network_operation(ns_lcm_op_occ_id)
cancel_sol_network_operation(ns_lcm_op_occ_id, params::Dict{String,<:Any})

Cancels a network operation. A network operation is any operation that is done to your network, such as network instance instantiation or termination.

Arguments

ns_lcm_op_occ_id: The identifier of the network operation.

Main.Tnb.create_sol_function_package — Method

create_sol_function_package()
create_sol_function_package(params::Dict{String,<:Any})

Creates a function package. A function package is a .zip file in CSAR (Cloud Service Archive) format that contains a network function (an ETSI standard telecommunication application) and function package descriptor that uses the TOSCA standard to describe how the network functions should run on your network. For more information, see Function packages in the Amazon Web Services Telco Network Builder User Guide. Creating a function package is the first step for creating a network in AWS TNB. This request creates an empty container with an ID. The next step is to upload the actual CSAR zip file into that empty container. To upload function package content, see PutSolFunctionPackageContent.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"tags": A tag is a label that you assign to an Amazon Web Services resource. Each tag consists of a key and an optional value. You can use tags to search and filter your resources or track your Amazon Web Services costs.

Main.Tnb.create_sol_network_instance — Method

create_sol_network_instance(ns_name, nsd_info_id)
create_sol_network_instance(ns_name, nsd_info_id, params::Dict{String,<:Any})

Creates a network instance. A network instance is a single network created in Amazon Web Services TNB that can be deployed and on which life-cycle operations (like terminate, update, and delete) can be performed. Creating a network instance is the third step after creating a network package. For more information about network instances, Network instances in the Amazon Web Services Telco Network Builder User Guide. Once you create a network instance, you can instantiate it. To instantiate a network, see InstantiateSolNetworkInstance.

Arguments

ns_name: Network instance name.

nsd_info_id: ID for network service descriptor.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nsDescription": Network instance description.

"tags": A tag is a label that you assign to an Amazon Web Services resource. Each tag consists of a key and an optional value. You can use tags to search and filter your resources or track your Amazon Web Services costs.

Main.Tnb.create_sol_network_package — Method

create_sol_network_package()
create_sol_network_package(params::Dict{String,<:Any})

Creates a network package. A network package is a .zip file in CSAR (Cloud Service Archive) format defines the function packages you want to deploy and the Amazon Web Services infrastructure you want to deploy them on. For more information, see Network instances in the Amazon Web Services Telco Network Builder User Guide. A network package consists of a network service descriptor (NSD) file (required) and any additional files (optional), such as scripts specific to your needs. For example, if you have multiple function packages in your network package, you can use the NSD to define which network functions should run in certain VPCs, subnets, or EKS clusters. This request creates an empty network package container with an ID. Once you create a network package, you can upload the network package content using PutSolNetworkPackageContent.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"tags": A tag is a label that you assign to an Amazon Web Services resource. Each tag consists of a key and an optional value. You can use tags to search and filter your resources or track your Amazon Web Services costs.

Main.Tnb.delete_sol_function_package — Method

delete_sol_function_package(vnf_pkg_id)
delete_sol_function_package(vnf_pkg_id, params::Dict{String,<:Any})

Deletes a function package. A function package is a .zip file in CSAR (Cloud Service Archive) format that contains a network function (an ETSI standard telecommunication application) and function package descriptor that uses the TOSCA standard to describe how the network functions should run on your network. To delete a function package, the package must be in a disabled state. To disable a function package, see UpdateSolFunctionPackage.

Arguments

vnf_pkg_id: ID of the function package.

Main.Tnb.delete_sol_network_instance — Method

delete_sol_network_instance(ns_instance_id)
delete_sol_network_instance(ns_instance_id, params::Dict{String,<:Any})

Deletes a network instance. A network instance is a single network created in Amazon Web Services TNB that can be deployed and on which life-cycle operations (like terminate, update, and delete) can be performed. To delete a network instance, the instance must be in a stopped or terminated state. To terminate a network instance, see TerminateSolNetworkInstance.

Arguments

ns_instance_id: Network instance ID.

Main.Tnb.delete_sol_network_package — Method

delete_sol_network_package(nsd_info_id)
delete_sol_network_package(nsd_info_id, params::Dict{String,<:Any})

Deletes network package. A network package is a .zip file in CSAR (Cloud Service Archive) format defines the function packages you want to deploy and the Amazon Web Services infrastructure you want to deploy them on. To delete a network package, the package must be in a disable state. To disable a network package, see UpdateSolNetworkPackage.

Arguments

nsd_info_id: ID of the network service descriptor in the network package.

Main.Tnb.get_sol_function_instance — Method

get_sol_function_instance(vnf_instance_id)
get_sol_function_instance(vnf_instance_id, params::Dict{String,<:Any})

Gets the details of a network function instance, including the instantation state and metadata from the function package descriptor in the network function package. A network function instance is a function in a function package .

Arguments

vnf_instance_id: ID of the network function.

Main.Tnb.get_sol_function_package — Method

get_sol_function_package(vnf_pkg_id)
get_sol_function_package(vnf_pkg_id, params::Dict{String,<:Any})

Gets the details of an individual function package, such as the operational state and whether the package is in use. A function package is a .zip file in CSAR (Cloud Service Archive) format that contains a network function (an ETSI standard telecommunication application) and function package descriptor that uses the TOSCA standard to describe how the network functions should run on your network..

Arguments

vnf_pkg_id: ID of the function package.

Main.Tnb.get_sol_function_package_content — Method

get_sol_function_package_content(accept, vnf_pkg_id)
get_sol_function_package_content(accept, vnf_pkg_id, params::Dict{String,<:Any})

Gets the contents of a function package. A function package is a .zip file in CSAR (Cloud Service Archive) format that contains a network function (an ETSI standard telecommunication application) and function package descriptor that uses the TOSCA standard to describe how the network functions should run on your network.

Arguments

accept: The format of the package that you want to download from the function packages.

vnf_pkg_id: ID of the function package.

Main.Tnb.get_sol_function_package_descriptor — Method

get_sol_function_package_descriptor(accept, vnf_pkg_id)
get_sol_function_package_descriptor(accept, vnf_pkg_id, params::Dict{String,<:Any})

Gets a function package descriptor in a function package. A function package descriptor is a .yaml file in a function package that uses the TOSCA standard to describe how the network function in the function package should run on your network. A function package is a .zip file in CSAR (Cloud Service Archive) format that contains a network function (an ETSI standard telecommunication application) and function package descriptor that uses the TOSCA standard to describe how the network functions should run on your network.

Arguments

accept: Indicates which content types, expressed as MIME types, the client is able to understand.

vnf_pkg_id: ID of the function package.

Main.Tnb.get_sol_network_instance — Method

get_sol_network_instance(ns_instance_id)
get_sol_network_instance(ns_instance_id, params::Dict{String,<:Any})

Gets the details of the network instance. A network instance is a single network created in Amazon Web Services TNB that can be deployed and on which life-cycle operations (like terminate, update, and delete) can be performed.

Arguments

ns_instance_id: ID of the network instance.

Main.Tnb.get_sol_network_operation — Method

get_sol_network_operation(ns_lcm_op_occ_id)
get_sol_network_operation(ns_lcm_op_occ_id, params::Dict{String,<:Any})

Gets the details of a network operation, including the tasks involved in the network operation and the status of the tasks. A network operation is any operation that is done to your network, such as network instance instantiation or termination.

Arguments

ns_lcm_op_occ_id: The identifier of the network operation.

Main.Tnb.get_sol_network_package — Method

get_sol_network_package(nsd_info_id)
get_sol_network_package(nsd_info_id, params::Dict{String,<:Any})

Gets the details of a network package. A network package is a .zip file in CSAR (Cloud Service Archive) format defines the function packages you want to deploy and the Amazon Web Services infrastructure you want to deploy them on.

Arguments

nsd_info_id: ID of the network service descriptor in the network package.

Main.Tnb.get_sol_network_package_content — Method

get_sol_network_package_content(accept, nsd_info_id)
get_sol_network_package_content(accept, nsd_info_id, params::Dict{String,<:Any})

Gets the contents of a network package. A network package is a .zip file in CSAR (Cloud Service Archive) format defines the function packages you want to deploy and the Amazon Web Services infrastructure you want to deploy them on.

Arguments

accept: The format of the package you want to download from the network package.

nsd_info_id: ID of the network service descriptor in the network package.

Main.Tnb.get_sol_network_package_descriptor — Method

get_sol_network_package_descriptor(nsd_info_id)
get_sol_network_package_descriptor(nsd_info_id, params::Dict{String,<:Any})

Gets the content of the network service descriptor. A network service descriptor is a .yaml file in a network package that uses the TOSCA standard to describe the network functions you want to deploy and the Amazon Web Services infrastructure you want to deploy the network functions on.

Arguments

nsd_info_id: ID of the network service descriptor in the network package.

Main.Tnb.instantiate_sol_network_instance — Method

instantiate_sol_network_instance(ns_instance_id)
instantiate_sol_network_instance(ns_instance_id, params::Dict{String,<:Any})

Instantiates a network instance. A network instance is a single network created in Amazon Web Services TNB that can be deployed and on which life-cycle operations (like terminate, update, and delete) can be performed. Before you can instantiate a network instance, you have to create a network instance. For more information, see CreateSolNetworkInstance.

Arguments

ns_instance_id: ID of the network instance.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"additionalParamsForNs": Provides values for the configurable properties.

"dry_run": A check for whether you have the required permissions for the action without actually making the request and provides an error response. If you have the required permissions, the error response is DryRunOperation. Otherwise, it is UnauthorizedOperation.

"tags": A tag is a label that you assign to an Amazon Web Services resource. Each tag consists of a key and an optional value. When you use this API, the tags are transferred to the network operation that is created. Use tags to search and filter your resources or track your Amazon Web Services costs.

Main.Tnb.list_sol_function_instances — Method

list_sol_function_instances()
list_sol_function_instances(params::Dict{String,<:Any})

Lists network function instances. A network function instance is a function in a function package .

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"max_results": The maximum number of results to include in the response.

"nextpage_opaque_marker": The token for the next page of results.

Main.Tnb.list_sol_function_packages — Method

list_sol_function_packages()
list_sol_function_packages(params::Dict{String,<:Any})

Lists information about function packages. A function package is a .zip file in CSAR (Cloud Service Archive) format that contains a network function (an ETSI standard telecommunication application) and function package descriptor that uses the TOSCA standard to describe how the network functions should run on your network.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"max_results": The maximum number of results to include in the response.

"nextpage_opaque_marker": The token for the next page of results.

Main.Tnb.list_sol_network_instances — Method

list_sol_network_instances()
list_sol_network_instances(params::Dict{String,<:Any})

Lists your network instances. A network instance is a single network created in Amazon Web Services TNB that can be deployed and on which life-cycle operations (like terminate, update, and delete) can be performed.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"max_results": The maximum number of results to include in the response.

"nextpage_opaque_marker": The token for the next page of results.

Main.Tnb.list_sol_network_operations — Method

list_sol_network_operations()
list_sol_network_operations(params::Dict{String,<:Any})

Lists details for a network operation, including when the operation started and the status of the operation. A network operation is any operation that is done to your network, such as network instance instantiation or termination.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"max_results": The maximum number of results to include in the response.

"nextpage_opaque_marker": The token for the next page of results.

Main.Tnb.list_sol_network_packages — Method

list_sol_network_packages()
list_sol_network_packages(params::Dict{String,<:Any})

Lists network packages. A network package is a .zip file in CSAR (Cloud Service Archive) format defines the function packages you want to deploy and the Amazon Web Services infrastructure you want to deploy them on.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"max_results": The maximum number of results to include in the response.

"nextpage_opaque_marker": The token for the next page of results.

Main.Tnb.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists tags for AWS TNB resources.

Arguments

resource_arn: Resource ARN.

Main.Tnb.put_sol_function_package_content — Method

put_sol_function_package_content(file, vnf_pkg_id)
put_sol_function_package_content(file, vnf_pkg_id, params::Dict{String,<:Any})

Uploads the contents of a function package. A function package is a .zip file in CSAR (Cloud Service Archive) format that contains a network function (an ETSI standard telecommunication application) and function package descriptor that uses the TOSCA standard to describe how the network functions should run on your network.

Arguments

file: Function package file.

vnf_pkg_id: Function package ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Content-Type": Function package content type.

Main.Tnb.put_sol_network_package_content — Method

put_sol_network_package_content(file, nsd_info_id)
put_sol_network_package_content(file, nsd_info_id, params::Dict{String,<:Any})

Uploads the contents of a network package. A network package is a .zip file in CSAR (Cloud Service Archive) format defines the function packages you want to deploy and the Amazon Web Services infrastructure you want to deploy them on.

Arguments

file: Network package file.

nsd_info_id: Network service descriptor info ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Content-Type": Network package content type.

Main.Tnb.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Tags an AWS TNB resource. A tag is a label that you assign to an Amazon Web Services resource. Each tag consists of a key and an optional value. You can use tags to search and filter your resources or track your Amazon Web Services costs.

Arguments

resource_arn: Resource ARN.

tags: A tag is a label that you assign to an Amazon Web Services resource. Each tag consists of a key and an optional value. You can use tags to search and filter your resources or track your Amazon Web Services costs.

Main.Tnb.terminate_sol_network_instance — Method

terminate_sol_network_instance(ns_instance_id)
terminate_sol_network_instance(ns_instance_id, params::Dict{String,<:Any})

Terminates a network instance. A network instance is a single network created in Amazon Web Services TNB that can be deployed and on which life-cycle operations (like terminate, update, and delete) can be performed. You must terminate a network instance before you can delete it.

Arguments

ns_instance_id: ID of the network instance.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"tags": A tag is a label that you assign to an Amazon Web Services resource. Each tag consists of a key and an optional value. When you use this API, the tags are transferred to the network operation that is created. Use tags to search and filter your resources or track your Amazon Web Services costs.

Main.Tnb.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Untags an AWS TNB resource. A tag is a label that you assign to an Amazon Web Services resource. Each tag consists of a key and an optional value. You can use tags to search and filter your resources or track your Amazon Web Services costs.

Arguments

resource_arn: Resource ARN.

tag_keys: Tag keys.

Main.Tnb.update_sol_function_package — Method

update_sol_function_package(operational_state, vnf_pkg_id)
update_sol_function_package(operational_state, vnf_pkg_id, params::Dict{String,<:Any})

Updates the operational state of function package. A function package is a .zip file in CSAR (Cloud Service Archive) format that contains a network function (an ETSI standard telecommunication application) and function package descriptor that uses the TOSCA standard to describe how the network functions should run on your network.

Arguments

operational_state: Operational state of the function package.

vnf_pkg_id: ID of the function package.

Main.Tnb.update_sol_network_instance — Method

update_sol_network_instance(ns_instance_id, update_type)
update_sol_network_instance(ns_instance_id, update_type, params::Dict{String,<:Any})

Update a network instance. A network instance is a single network created in Amazon Web Services TNB that can be deployed and on which life-cycle operations (like terminate, update, and delete) can be performed.

Arguments

ns_instance_id: ID of the network instance.

update_type: The type of update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"modifyVnfInfoData": Identifies the network function information parameters and/or the configurable properties of the network function to be modified.

"tags": A tag is a label that you assign to an Amazon Web Services resource. Each tag consists of a key and an optional value. When you use this API, the tags are transferred to the network operation that is created. Use tags to search and filter your resources or track your Amazon Web Services costs.

Main.Tnb.update_sol_network_package — Method

update_sol_network_package(nsd_info_id, nsd_operational_state)
update_sol_network_package(nsd_info_id, nsd_operational_state, params::Dict{String,<:Any})

Updates the operational state of a network package. A network package is a .zip file in CSAR (Cloud Service Archive) format defines the function packages you want to deploy and the Amazon Web Services infrastructure you want to deploy them on. A network service descriptor is a .yaml file in a network package that uses the TOSCA standard to describe the network functions you want to deploy and the Amazon Web Services infrastructure you want to deploy the network functions on.

Arguments

nsd_info_id: ID of the network service descriptor in the network package.

nsd_operational_state: Operational state of the network service descriptor in the network package.

Main.Tnb.validate_sol_function_package_content — Method

validate_sol_function_package_content(file, vnf_pkg_id)
validate_sol_function_package_content(file, vnf_pkg_id, params::Dict{String,<:Any})

Validates function package content. This can be used as a dry run before uploading function package content with PutSolFunctionPackageContent. A function package is a .zip file in CSAR (Cloud Service Archive) format that contains a network function (an ETSI standard telecommunication application) and function package descriptor that uses the TOSCA standard to describe how the network functions should run on your network.

Arguments

file: Function package file.

vnf_pkg_id: Function package ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Content-Type": Function package content type.

Main.Tnb.validate_sol_network_package_content — Method

validate_sol_network_package_content(file, nsd_info_id)
validate_sol_network_package_content(file, nsd_info_id, params::Dict{String,<:Any})

Validates network package content. This can be used as a dry run before uploading network package content with PutSolNetworkPackageContent. A network package is a .zip file in CSAR (Cloud Service Archive) format defines the function packages you want to deploy and the Amazon Web Services infrastructure you want to deploy them on.

Arguments

file: Network package file.

nsd_info_id: Network service descriptor file.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Content-Type": Network package content type.