Transcribe Streaming

This page documents function available when using the Transcribe_Streaming module, created with @service Transcribe_Streaming.

Index

Main.Transcribe_Streaming.start_call_analytics_stream_transcription

Main.Transcribe_Streaming.start_medical_stream_transcription

Main.Transcribe_Streaming.start_stream_transcription

Documentation

Main.Transcribe_Streaming.start_call_analytics_stream_transcription — Function

start_call_analytics_stream_transcription(audio_stream, x-amzn-transcribe-language-code, x-amzn-transcribe-media-encoding, x-amzn-transcribe-sample-rate)
start_call_analytics_stream_transcription(audio_stream, x-amzn-transcribe-language-code, x-amzn-transcribe-media-encoding, x-amzn-transcribe-sample-rate, params::Dict{String,<:Any})

Starts a bidirectional HTTP/2 or WebSocket stream where audio is streamed to Amazon Transcribe and the transcription results are streamed to your application. Use this operation for Call Analytics transcriptions.

The following parameters are required:

language-code

media-encoding

sample-rate

For more information on streaming with Amazon Transcribe, see Transcribing streaming audio.

Arguments

audio_stream:

x-amzn-transcribe-language-code: Specify the language code that represents the language spoken in your audio.

If you're unsure of the language spoken in your audio, consider using IdentifyLanguage to enable automatic language identification.

For a list of languages supported with streaming Call Analytics, refer to the Supported languages table.

x-amzn-transcribe-media-encoding: Specify the encoding of your input audio. Supported formats are:

FLAC

OPUS-encoded audio in an Ogg container

PCM (only signed 16-bit little-endian audio formats, which does not include WAV)

For more information, see Media formats.

x-amzn-transcribe-sample-rate: The sample rate of the input audio (in hertz). Low- quality audio, such as telephone audio, is typically around 8,000 Hz. High-quality audio typically ranges from 16,000 Hz to 48,000 Hz. Note that the sample rate you specify must match that of your audio.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"x-amzn-transcribe-content-identification-type": Labels all personally identifiable information (PII) identified in your transcript.

Content identification is performed at the segment level; PII specified in PiiEntityTypes is flagged upon complete transcription of an audio segment.

You can’t set ContentIdentificationType and ContentRedactionType in the same request. If you set both, your request returns a BadRequestException.

For more information, see Redacting or identifying personally identifiable information.

"x-amzn-transcribe-content-redaction-type": Redacts all personally identifiable information (PII) identified in your transcript.

Content redaction is performed at the segment level; PII specified in PiiEntityTypes is redacted upon complete transcription of an audio segment.

You can’t set ContentRedactionType and ContentIdentificationType in the same request. If you set both, your request returns a BadRequestException.

For more information, see Redacting or identifying personally identifiable information.

"x-amzn-transcribe-enable-partial-results-stabilization": Enables partial result stabilization for your transcription. Partial result stabilization can reduce latency in your output, but may impact accuracy. For more information, see Partial-result stabilization.

"x-amzn-transcribe-language-model-name": Specify the name of the custom language model that you want to use when processing your transcription. Note that language model names are case sensitive.

The language of the specified language model must match the language code you specify in your transcription request. If the languages don't match, the custom language model isn't applied. There are no errors or warnings associated with a language mismatch.

For more information, see Custom language models.

"x-amzn-transcribe-partial-results-stability": Specify the level of stability to use when you enable partial results stabilization (EnablePartialResultsStabilization).

Low stability provides the highest accuracy. High stability transcribes faster, but with slightly lower accuracy.

For more information, see Partial-result stabilization.

"x-amzn-transcribe-pii-entity-types": Specify which types of personally identifiable information (PII) you want to redact in your transcript. You can include as many types as you'd like, or you can select ALL.

To include PiiEntityTypes in your Call Analytics request, you must also include either ContentIdentificationType or ContentRedactionType.

Values must be comma-separated and can include: BANK_ACCOUNT_NUMBER, BANK_ROUTING, CREDIT_DEBIT_NUMBER, CREDIT_DEBIT_CVV, CREDIT_DEBIT_EXPIRY, PIN, EMAIL, ADDRESS, NAME, PHONE, SSN, or ALL.

"x-amzn-transcribe-session-id": Specify a name for your Call Analytics transcription session. If you don't include this parameter in your request, Amazon Transcribe generates an ID and returns it in the response.

You can use a session ID to retry a streaming session.

"x-amzn-transcribe-vocabulary-filter-method": Specify how you want your vocabulary filter applied to your transcript.

To replace words with ***, choose mask.

To delete words, choose remove.

To flag words without changing them, choose tag.

"x-amzn-transcribe-vocabulary-filter-name": Specify the name of the custom vocabulary filter that you want to use when processing your transcription. Note that vocabulary filter names are case sensitive.

If the language of the specified custom vocabulary filter doesn't match the language identified in your media, the vocabulary filter is not applied to your transcription.

For more information, see Using vocabulary filtering with unwanted words.

"x-amzn-transcribe-vocabulary-name": Specify the name of the custom vocabulary that you want to use when processing your transcription. Note that vocabulary names are case sensitive.

If the language of the specified custom vocabulary doesn't match the language identified in your media, the custom vocabulary is not applied to your transcription.

For more information, see Custom vocabularies.

Main.Transcribe_Streaming.start_medical_stream_transcription — Function

start_medical_stream_transcription(audio_stream, x-amzn-transcribe-language-code, x-amzn-transcribe-media-encoding, x-amzn-transcribe-sample-rate, x-amzn-transcribe-specialty, x-amzn-transcribe-type)
start_medical_stream_transcription(audio_stream, x-amzn-transcribe-language-code, x-amzn-transcribe-media-encoding, x-amzn-transcribe-sample-rate, x-amzn-transcribe-specialty, x-amzn-transcribe-type, params::Dict{String,<:Any})

Starts a bidirectional HTTP/2 or WebSocket stream where audio is streamed to Amazon Transcribe Medical and the transcription results are streamed to your application.

The following parameters are required:

language-code

media-encoding

sample-rate

For more information on streaming with Amazon Transcribe Medical, see Transcribing streaming audio.

Arguments

audio_stream:

x-amzn-transcribe-language-code: Specify the language code that represents the language spoken in your audio.

Important

Amazon Transcribe Medical only supports US English (en-US).

x-amzn-transcribe-media-encoding: Specify the encoding used for the input audio. Supported formats are:

FLAC

OPUS-encoded audio in an Ogg container

PCM (only signed 16-bit little-endian audio formats, which does not include WAV)

For more information, see Media formats.

x-amzn-transcribe-sample-rate: The sample rate of the input audio (in hertz). Amazon Transcribe Medical supports a range from 16,000 Hz to 48,000 Hz. Note that the sample rate you specify must match that of your audio.

x-amzn-transcribe-specialty: Specify the medical specialty contained in your audio.

x-amzn-transcribe-type: Specify the type of input audio. For example, choose DICTATION for a provider dictating patient notes and CONVERSATION for a dialogue between a patient and a medical professional.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"x-amzn-transcribe-content-identification-type": Labels all personal health information (PHI) identified in your transcript.

Content identification is performed at the segment level; PHI is flagged upon complete transcription of an audio segment.

For more information, see Identifying personal health information (PHI) in a transcription.

"x-amzn-transcribe-enable-channel-identification": Enables channel identification in multi-channel audio.

Channel identification transcribes the audio on each channel independently, then appends the output for each channel into one transcript.

If you have multi-channel audio and do not enable channel identification, your audio is transcribed in a continuous manner and your transcript is not separated by channel.

For more information, see Transcribing multi-channel audio.

"x-amzn-transcribe-number-of-channels": Specify the number of channels in your audio stream. Up to two channels are supported.

"x-amzn-transcribe-session-id": Specify a name for your transcription session. If you don't include this parameter in your request, Amazon Transcribe Medical generates an ID and returns it in the response.

You can use a session ID to retry a streaming session.

"x-amzn-transcribe-show-speaker-label": Enables speaker partitioning (diarization) in your transcription output. Speaker partitioning labels the speech from individual speakers in your media file.

For more information, see Partitioning speakers (diarization).

"x-amzn-transcribe-vocabulary-name": Specify the name of the custom vocabulary that you want to use when processing your transcription. Note that vocabulary names are case sensitive.

Main.Transcribe_Streaming.start_stream_transcription — Function

start_stream_transcription(audio_stream, x-amzn-transcribe-media-encoding, x-amzn-transcribe-sample-rate)
start_stream_transcription(audio_stream, x-amzn-transcribe-media-encoding, x-amzn-transcribe-sample-rate, params::Dict{String,<:Any})

Starts a bidirectional HTTP/2 or WebSocket stream where audio is streamed to Amazon Transcribe and the transcription results are streamed to your application.

The following parameters are required:

language-code or identify-language or identify-multiple-language

media-encoding

sample-rate

For more information on streaming with Amazon Transcribe, see Transcribing streaming audio.

Arguments

audio_stream: An encoded stream of audio blobs. Audio streams are encoded as either HTTP/2 or WebSocket data frames.

For more information, see Transcribing streaming audio.

x-amzn-transcribe-media-encoding: Specify the encoding of your input audio. Supported formats are:

FLAC

OPUS-encoded audio in an Ogg container

PCM (only signed 16-bit little-endian audio formats, which does not include WAV)

For more information, see Media formats.

x-amzn-transcribe-sample-rate: The sample rate of the input audio (in hertz). Low- quality audio, such as telephone audio, is typically around 8,000 Hz. High-quality audio typically ranges from 16,000 Hz to 48,000 Hz. Note that the sample rate you specify must match that of your audio.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"x-amzn-transcribe-content-identification-type": Labels all personally identifiable information (PII) identified in your transcript.

Content identification is performed at the segment level; PII specified in PiiEntityTypes is flagged upon complete transcription of an audio segment.

You can’t set ContentIdentificationType and ContentRedactionType in the same request. If you set both, your request returns a BadRequestException.

For more information, see Redacting or identifying personally identifiable information.

"x-amzn-transcribe-content-redaction-type": Redacts all personally identifiable information (PII) identified in your transcript.

Content redaction is performed at the segment level; PII specified in PiiEntityTypes is redacted upon complete transcription of an audio segment.

You can’t set ContentRedactionType and ContentIdentificationType in the same request. If you set both, your request returns a BadRequestException.

For more information, see Redacting or identifying personally identifiable information.

"x-amzn-transcribe-enable-channel-identification": Enables channel identification in multi-channel audio.

Channel identification transcribes the audio on each channel independently, then appends the output for each channel into one transcript.

If you have multi-channel audio and do not enable channel identification, your audio is transcribed in a continuous manner and your transcript is not separated by channel.

For more information, see Transcribing multi-channel audio.

"x-amzn-transcribe-enable-partial-results-stabilization": Enables partial result stabilization for your transcription. Partial result stabilization can reduce latency in your output, but may impact accuracy. For more information, see Partial-result stabilization.

"x-amzn-transcribe-identify-language": Enables automatic language identification for your transcription.

If you include IdentifyLanguage, you can optionally include a list of language codes, using LanguageOptions, that you think may be present in your audio stream. Including language options can improve transcription accuracy.

You can also include a preferred language using PreferredLanguage. Adding a preferred language can help Amazon Transcribe identify the language faster than if you omit this parameter.

If you have multi-channel audio that contains different languages on each channel, and you've enabled channel identification, automatic language identification identifies the dominant language on each audio channel.

Note that you must include either LanguageCode or IdentifyLanguage or IdentifyMultipleLanguages in your request. If you include more than one of these parameters, your transcription job fails.

Streaming language identification can't be combined with custom language models or redaction.

"x-amzn-transcribe-identify-multiple-languages": Enables automatic multi-language identification in your transcription job request. Use this parameter if your stream contains more than one language. If your stream contains only one language, use IdentifyLanguage instead.

If you include IdentifyMultipleLanguages, you can optionally include a list of language codes, using LanguageOptions, that you think may be present in your stream. Including LanguageOptions restricts IdentifyMultipleLanguages to only the language options that you specify, which can improve transcription accuracy.

If you want to apply a custom vocabulary or a custom vocabulary filter to your automatic multiple language identification request, include VocabularyNames or VocabularyFilterNames.

Note that you must include one of LanguageCode, IdentifyLanguage, or IdentifyMultipleLanguages in your request. If you include more than one of these parameters, your transcription job fails.

"x-amzn-transcribe-language-code": Specify the language code that represents the language spoken in your audio.

If you're unsure of the language spoken in your audio, consider using IdentifyLanguage to enable automatic language identification.

For a list of languages supported with Amazon Transcribe streaming, refer to the Supported languages table.

"x-amzn-transcribe-language-model-name": Specify the name of the custom language model that you want to use when processing your transcription. Note that language model names are case sensitive.

The language of the specified language model must match the language code you specify in your transcription request. If the languages don't match, the custom language model isn't applied. There are no errors or warnings associated with a language mismatch.

For more information, see Custom language models.

"x-amzn-transcribe-language-options": Specify two or more language codes that represent the languages you think may be present in your media; including more than five is not recommended. If you're unsure what languages are present, do not include this parameter.

Including language options can improve the accuracy of language identification.

If you include LanguageOptions in your request, you must also include IdentifyLanguage.

For a list of languages supported with Amazon Transcribe streaming, refer to the Supported languages table.

Important

You can only include one language dialect per language per stream. For example, you cannot include en-US and en-AU in the same request.

"x-amzn-transcribe-number-of-channels": Specify the number of channels in your audio stream. Up to two channels are supported.

"x-amzn-transcribe-partial-results-stability": Specify the level of stability to use when you enable partial results stabilization (EnablePartialResultsStabilization).

Low stability provides the highest accuracy. High stability transcribes faster, but with slightly lower accuracy.

For more information, see Partial-result stabilization.

"x-amzn-transcribe-pii-entity-types": Specify which types of personally identifiable information (PII) you want to redact in your transcript. You can include as many types as you'd like, or you can select ALL.

To include PiiEntityTypes in your request, you must also include either ContentIdentificationType or ContentRedactionType.

Values must be comma-separated and can include: BANK_ACCOUNT_NUMBER, BANK_ROUTING, CREDIT_DEBIT_NUMBER, CREDIT_DEBIT_CVV, CREDIT_DEBIT_EXPIRY, PIN, EMAIL, ADDRESS, NAME, PHONE, SSN, or ALL.

"x-amzn-transcribe-preferred-language": Specify a preferred language from the subset of languages codes you specified in LanguageOptions.

You can only use this parameter if you've included IdentifyLanguage and LanguageOptions in your request.

"x-amzn-transcribe-session-id": Specify a name for your transcription session. If you don't include this parameter in your request, Amazon Transcribe generates an ID and returns it in the response.

You can use a session ID to retry a streaming session.

"x-amzn-transcribe-show-speaker-label": Enables speaker partitioning (diarization) in your transcription output. Speaker partitioning labels the speech from individual speakers in your media file.

For more information, see Partitioning speakers (diarization).

"x-amzn-transcribe-vocabulary-filter-method": Specify how you want your vocabulary filter applied to your transcript.

To replace words with ***, choose mask.

To delete words, choose remove.

To flag words without changing them, choose tag.

"x-amzn-transcribe-vocabulary-filter-name": Specify the name of the custom vocabulary filter that you want to use when processing your transcription. Note that vocabulary filter names are case sensitive.

If the language of the specified custom vocabulary filter doesn't match the language identified in your media, the vocabulary filter is not applied to your transcription.

Important

This parameter is not intended for use with the IdentifyLanguage parameter. If you're including IdentifyLanguage in your request and want to use one or more vocabulary filters with your transcription, use the VocabularyFilterNames parameter instead.

For more information, see Using vocabulary filtering with unwanted words.

"x-amzn-transcribe-vocabulary-filter-names": Specify the names of the custom vocabulary filters that you want to use when processing your transcription. Note that vocabulary filter names are case sensitive.

If none of the languages of the specified custom vocabulary filters match the language identified in your media, your job fails.

Important

This parameter is only intended for use with the IdentifyLanguage parameter. If you're not including IdentifyLanguage in your request and want to use a custom vocabulary filter with your transcription, use the VocabularyFilterName parameter instead.

For more information, see Using vocabulary filtering with unwanted words.

"x-amzn-transcribe-vocabulary-name": Specify the name of the custom vocabulary that you want to use when processing your transcription. Note that vocabulary names are case sensitive.

If the language of the specified custom vocabulary doesn't match the language identified in your media, the custom vocabulary is not applied to your transcription.

Important

This parameter is not intended for use with the IdentifyLanguage parameter. If you're including IdentifyLanguage in your request and want to use one or more custom vocabularies with your transcription, use the VocabularyNames parameter instead.

For more information, see Custom vocabularies.

"x-amzn-transcribe-vocabulary-names": Specify the names of the custom vocabularies that you want to use when processing your transcription. Note that vocabulary names are case sensitive.

If none of the languages of the specified custom vocabularies match the language identified in your media, your job fails.

Important

This parameter is only intended for use with the IdentifyLanguage parameter. If you're not including IdentifyLanguage in your request and want to use a custom vocabulary with your transcription, use the VocabularyName parameter instead.

For more information, see Custom vocabularies.