Voice Id

This page documents function available when using the Voice_Id module, created with @service Voice_Id.

Index

Main.Voice_Id.associate_fraudster

Main.Voice_Id.create_domain

Main.Voice_Id.create_watchlist

Main.Voice_Id.delete_domain

Main.Voice_Id.delete_fraudster

Main.Voice_Id.delete_speaker

Main.Voice_Id.delete_watchlist

Main.Voice_Id.describe_domain

Main.Voice_Id.describe_fraudster

Main.Voice_Id.describe_fraudster_registration_job

Main.Voice_Id.describe_speaker

Main.Voice_Id.describe_speaker_enrollment_job

Main.Voice_Id.describe_watchlist

Main.Voice_Id.disassociate_fraudster

Main.Voice_Id.evaluate_session

Main.Voice_Id.list_domains

Main.Voice_Id.list_fraudster_registration_jobs

Main.Voice_Id.list_fraudsters

Main.Voice_Id.list_speaker_enrollment_jobs

Main.Voice_Id.list_speakers

Main.Voice_Id.list_tags_for_resource

Main.Voice_Id.list_watchlists

Main.Voice_Id.opt_out_speaker

Main.Voice_Id.start_fraudster_registration_job

Main.Voice_Id.start_speaker_enrollment_job

Main.Voice_Id.tag_resource

Main.Voice_Id.untag_resource

Main.Voice_Id.update_domain

Main.Voice_Id.update_watchlist

Documentation

Main.Voice_Id.associate_fraudster — Method

associate_fraudster(domain_id, fraudster_id, watchlist_id)
associate_fraudster(domain_id, fraudster_id, watchlist_id, params::Dict{String,<:Any})

Associates the fraudsters with the watchlist specified in the same domain.

Arguments

domain_id: The identifier of the domain that contains the fraudster.

fraudster_id: The identifier of the fraudster to be associated with the watchlist.

watchlist_id: The identifier of the watchlist you want to associate with the fraudster.

Main.Voice_Id.create_domain — Method

create_domain(name, server_side_encryption_configuration)
create_domain(name, server_side_encryption_configuration, params::Dict{String,<:Any})

Creates a domain that contains all Amazon Connect Voice ID data, such as speakers, fraudsters, customer audio, and voiceprints. Every domain is created with a default watchlist that fraudsters can be a part of.

Arguments

name: The name of the domain.

server_side_encryption_configuration: The configuration, containing the KMS key identifier, to be used by Voice ID for the server-side encryption of your data. Refer to Amazon Connect Voice ID encryption at rest for more details on how the KMS key is used.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If not provided, the Amazon Web Services SDK populates this field. For more information about idempotency, see Making retries safe with idempotent APIs.

"Description": A brief description of this domain.

"Tags": A list of tags you want added to the domain.

Main.Voice_Id.create_watchlist — Method

create_watchlist(domain_id, name)
create_watchlist(domain_id, name, params::Dict{String,<:Any})

Creates a watchlist that fraudsters can be a part of.

Arguments

domain_id: The identifier of the domain that contains the watchlist.

name: The name of the watchlist.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If not provided, the Amazon Web Services SDK populates this field. For more information about idempotency, see Making retries safe with idempotent APIs.

"Description": A brief description of this watchlist.

Main.Voice_Id.delete_domain — Method

delete_domain(domain_id)
delete_domain(domain_id, params::Dict{String,<:Any})

Deletes the specified domain from Voice ID.

Arguments

domain_id: The identifier of the domain you want to delete.

Main.Voice_Id.delete_fraudster — Method

delete_fraudster(domain_id, fraudster_id)
delete_fraudster(domain_id, fraudster_id, params::Dict{String,<:Any})

Deletes the specified fraudster from Voice ID. This action disassociates the fraudster from any watchlists it is a part of.

Arguments

domain_id: The identifier of the domain that contains the fraudster.

fraudster_id: The identifier of the fraudster you want to delete.

Main.Voice_Id.delete_speaker — Method

delete_speaker(domain_id, speaker_id)
delete_speaker(domain_id, speaker_id, params::Dict{String,<:Any})

Deletes the specified speaker from Voice ID.

Arguments

domain_id: The identifier of the domain that contains the speaker.

speaker_id: The identifier of the speaker you want to delete.

Main.Voice_Id.delete_watchlist — Method

delete_watchlist(domain_id, watchlist_id)
delete_watchlist(domain_id, watchlist_id, params::Dict{String,<:Any})

Deletes the specified watchlist from Voice ID. This API throws an exception when there are fraudsters in the watchlist that you are trying to delete. You must delete the fraudsters, and then delete the watchlist. Every domain has a default watchlist which cannot be deleted.

Arguments

domain_id: The identifier of the domain that contains the watchlist.

watchlist_id: The identifier of the watchlist to be deleted.

Main.Voice_Id.describe_domain — Method

describe_domain(domain_id)
describe_domain(domain_id, params::Dict{String,<:Any})

Describes the specified domain.

Arguments

domain_id: The identifier of the domain that you are describing.

Main.Voice_Id.describe_fraudster — Method

describe_fraudster(domain_id, fraudster_id)
describe_fraudster(domain_id, fraudster_id, params::Dict{String,<:Any})

Describes the specified fraudster.

Arguments

domain_id: The identifier of the domain that contains the fraudster.

fraudster_id: The identifier of the fraudster you are describing.

Main.Voice_Id.describe_fraudster_registration_job — Method

describe_fraudster_registration_job(domain_id, job_id)
describe_fraudster_registration_job(domain_id, job_id, params::Dict{String,<:Any})

Describes the specified fraudster registration job.

Arguments

domain_id: The identifier of the domain that contains the fraudster registration job.

job_id: The identifier of the fraudster registration job you are describing.

Main.Voice_Id.describe_speaker — Method

describe_speaker(domain_id, speaker_id)
describe_speaker(domain_id, speaker_id, params::Dict{String,<:Any})

Describes the specified speaker.

Arguments

domain_id: The identifier of the domain that contains the speaker.

speaker_id: The identifier of the speaker you are describing.

Main.Voice_Id.describe_speaker_enrollment_job — Method

describe_speaker_enrollment_job(domain_id, job_id)
describe_speaker_enrollment_job(domain_id, job_id, params::Dict{String,<:Any})

Describes the specified speaker enrollment job.

Arguments

domain_id: The identifier of the domain that contains the speaker enrollment job.

job_id: The identifier of the speaker enrollment job you are describing.

Main.Voice_Id.describe_watchlist — Method

describe_watchlist(domain_id, watchlist_id)
describe_watchlist(domain_id, watchlist_id, params::Dict{String,<:Any})

Describes the specified watchlist.

Arguments

domain_id: The identifier of the domain that contains the watchlist.

watchlist_id: The identifier of the watchlist that you are describing.

Main.Voice_Id.disassociate_fraudster — Method

disassociate_fraudster(domain_id, fraudster_id, watchlist_id)
disassociate_fraudster(domain_id, fraudster_id, watchlist_id, params::Dict{String,<:Any})

Disassociates the fraudsters from the watchlist specified. Voice ID always expects a fraudster to be a part of at least one watchlist. If you try to disassociate a fraudster from its only watchlist, a ValidationException is thrown.

Arguments

domain_id: The identifier of the domain that contains the fraudster.

fraudster_id: The identifier of the fraudster to be disassociated from the watchlist.

watchlist_id: The identifier of the watchlist that you want to disassociate from the fraudster.

Main.Voice_Id.evaluate_session — Method

evaluate_session(domain_id, session_name_or_id)
evaluate_session(domain_id, session_name_or_id, params::Dict{String,<:Any})

Evaluates a specified session based on audio data accumulated during a streaming Amazon Connect Voice ID call.

Arguments

domain_id: The identifier of the domain where the session started.

session_name_or_id: The session identifier, or name of the session, that you want to evaluate. In Voice ID integration, this is the Contact-Id.

Main.Voice_Id.list_domains — Method

list_domains()
list_domains(params::Dict{String,<:Any})

Lists all the domains in the Amazon Web Services account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results that are returned per call. You can use NextToken to obtain more pages of results. The default is 100; the maximum allowed page size is also 100.

"NextToken": If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

Main.Voice_Id.list_fraudster_registration_jobs — Method

list_fraudster_registration_jobs(domain_id)
list_fraudster_registration_jobs(domain_id, params::Dict{String,<:Any})

Lists all the fraudster registration jobs in the domain with the given JobStatus. If JobStatus is not provided, this lists all fraudster registration jobs in the given domain.

Arguments

domain_id: The identifier of the domain that contains the fraudster registration Jobs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"JobStatus": Provides the status of your fraudster registration job.

"MaxResults": The maximum number of results that are returned per call. You can use NextToken to obtain more pages of results. The default is 100; the maximum allowed page size is also 100.

"NextToken": If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

Main.Voice_Id.list_fraudsters — Method

list_fraudsters(domain_id)
list_fraudsters(domain_id, params::Dict{String,<:Any})

Lists all fraudsters in a specified watchlist or domain.

Arguments

domain_id: The identifier of the domain.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results that are returned per call. You can use NextToken to obtain more pages of results. The default is 100; the maximum allowed page size is also 100.

"NextToken": If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

"WatchlistId": The identifier of the watchlist. If provided, all fraudsters in the watchlist are listed. If not provided, all fraudsters in the domain are listed.

Main.Voice_Id.list_speaker_enrollment_jobs — Method

list_speaker_enrollment_jobs(domain_id)
list_speaker_enrollment_jobs(domain_id, params::Dict{String,<:Any})

Lists all the speaker enrollment jobs in the domain with the specified JobStatus. If JobStatus is not provided, this lists all jobs with all possible speaker enrollment job statuses.

Arguments

domain_id: The identifier of the domain that contains the speaker enrollment jobs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"JobStatus": Provides the status of your speaker enrollment Job.

"MaxResults": The maximum number of results that are returned per call. You can use NextToken to obtain more pages of results. The default is 100; the maximum allowed page size is also 100.

"NextToken": If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

Main.Voice_Id.list_speakers — Method

list_speakers(domain_id)
list_speakers(domain_id, params::Dict{String,<:Any})

Lists all speakers in a specified domain.

Arguments

domain_id: The identifier of the domain.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results that are returned per call. You can use NextToken to obtain more pages of results. The default is 100; the maximum allowed page size is also 100.

"NextToken": If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

Main.Voice_Id.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists all tags associated with a specified Voice ID resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the Voice ID resource for which you want to list the tags.

Main.Voice_Id.list_watchlists — Method

list_watchlists(domain_id)
list_watchlists(domain_id, params::Dict{String,<:Any})

Lists all watchlists in a specified domain.

Arguments

domain_id: The identifier of the domain.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results that are returned per call. You can use NextToken to obtain more pages of results. The default is 100; the maximum allowed page size is also 100.

"NextToken": If NextToken is returned, there are more results available. The value of NextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours.

Main.Voice_Id.opt_out_speaker — Method

opt_out_speaker(domain_id, speaker_id)
opt_out_speaker(domain_id, speaker_id, params::Dict{String,<:Any})

Opts out a speaker from Voice ID. A speaker can be opted out regardless of whether or not they already exist in Voice ID. If they don't yet exist, a new speaker is created in an opted out state. If they already exist, their existing status is overridden and they are opted out. Enrollment and evaluation authentication requests are rejected for opted out speakers, and opted out speakers have no voice embeddings stored in Voice ID.

Arguments

domain_id: The identifier of the domain that contains the speaker.

speaker_id: The identifier of the speaker you want opted-out.

Main.Voice_Id.start_fraudster_registration_job — Method

start_fraudster_registration_job(data_access_role_arn, domain_id, input_data_config, output_data_config)
start_fraudster_registration_job(data_access_role_arn, domain_id, input_data_config, output_data_config, params::Dict{String,<:Any})

Starts a new batch fraudster registration job using provided details.

Arguments

data_access_role_arn: The IAM role Amazon Resource Name (ARN) that grants Voice ID permissions to access customer's buckets to read the input manifest file and write the Job output file. Refer to the Create and edit a fraudster watchlist documentation for the permissions needed in this role.

domain_id: The identifier of the domain that contains the fraudster registration job and in which the fraudsters are registered.

input_data_config: The input data config containing an S3 URI for the input manifest file that contains the list of fraudster registration requests.

output_data_config: The output data config containing the S3 location where Voice ID writes the job output file; you must also include a KMS key ID to encrypt the file.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If not provided, the Amazon Web Services SDK populates this field. For more information about idempotency, see Making retries safe with idempotent APIs.

"JobName": The name of the new fraudster registration job.

"RegistrationConfig": The registration config containing details such as the action to take when a duplicate fraudster is detected, and the similarity threshold to use for detecting a duplicate fraudster.

Main.Voice_Id.start_speaker_enrollment_job — Method

start_speaker_enrollment_job(data_access_role_arn, domain_id, input_data_config, output_data_config)
start_speaker_enrollment_job(data_access_role_arn, domain_id, input_data_config, output_data_config, params::Dict{String,<:Any})

Starts a new batch speaker enrollment job using specified details.

Arguments

data_access_role_arn: The IAM role Amazon Resource Name (ARN) that grants Voice ID permissions to access customer's buckets to read the input manifest file and write the job output file. Refer to Batch enrollment using audio data from prior calls for the permissions needed in this role.

domain_id: The identifier of the domain that contains the speaker enrollment job and in which the speakers are enrolled.

input_data_config: The input data config containing the S3 location for the input manifest file that contains the list of speaker enrollment requests.

output_data_config: The output data config containing the S3 location where Voice ID writes the job output file; you must also include a KMS key ID to encrypt the file.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. If not provided, the Amazon Web Services SDK populates this field. For more information about idempotency, see Making retries safe with idempotent APIs.

"EnrollmentConfig": The enrollment config that contains details such as the action to take when a speaker is already enrolled in Voice ID or when a speaker is identified as a fraudster.

"JobName": A name for your speaker enrollment job.

Main.Voice_Id.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Tags a Voice ID resource with the provided list of tags.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the Voice ID resource you want to tag.

tags: The list of tags to assign to the specified resource.

Main.Voice_Id.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes specified tags from a specified Amazon Connect Voice ID resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the Voice ID resource you want to remove tags from.

tag_keys: The list of tag keys you want to remove from the specified resource.

Main.Voice_Id.update_domain — Method

update_domain(domain_id, name, server_side_encryption_configuration)
update_domain(domain_id, name, server_side_encryption_configuration, params::Dict{String,<:Any})

Updates the specified domain. This API has clobber behavior, and clears and replaces all attributes. If an optional field, such as 'Description' is not provided, it is removed from the domain.

Arguments

domain_id: The identifier of the domain to be updated.

name: The name of the domain.

server_side_encryption_configuration: The configuration, containing the KMS key identifier, to be used by Voice ID for the server-side encryption of your data. Changing the domain's associated KMS key immediately triggers an asynchronous process to remove dependency on the old KMS key, such that the domain's data can only be accessed using the new KMS key. The domain's ServerSideEncryptionUpdateDetails contains the details for this process.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A brief description about this domain.

Main.Voice_Id.update_watchlist — Method

update_watchlist(domain_id, watchlist_id)
update_watchlist(domain_id, watchlist_id, params::Dict{String,<:Any})

Updates the specified watchlist. Every domain has a default watchlist which cannot be updated.

Arguments

domain_id: The identifier of the domain that contains the watchlist.

watchlist_id: The identifier of the watchlist to be updated.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A brief description about this watchlist.

"Name": The name of the watchlist.