Wafv2

This page documents function available when using the Wafv2 module, created with @service Wafv2.

Index

Main.Wafv2.associate_web_acl

Main.Wafv2.check_capacity

Main.Wafv2.create_ipset

Main.Wafv2.create_regex_pattern_set

Main.Wafv2.create_rule_group

Main.Wafv2.create_web_acl

Main.Wafv2.delete_firewall_manager_rule_groups

Main.Wafv2.delete_ipset

Main.Wafv2.delete_logging_configuration

Main.Wafv2.delete_permission_policy

Main.Wafv2.delete_regex_pattern_set

Main.Wafv2.delete_rule_group

Main.Wafv2.delete_web_acl

Main.Wafv2.describe_managed_rule_group

Main.Wafv2.disassociate_web_acl

Main.Wafv2.get_ipset

Main.Wafv2.get_logging_configuration

Main.Wafv2.get_permission_policy

Main.Wafv2.get_rate_based_statement_managed_keys

Main.Wafv2.get_regex_pattern_set

Main.Wafv2.get_rule_group

Main.Wafv2.get_sampled_requests

Main.Wafv2.get_web_acl

Main.Wafv2.get_web_aclfor_resource

Main.Wafv2.list_available_managed_rule_groups

Main.Wafv2.list_ipsets

Main.Wafv2.list_logging_configurations

Main.Wafv2.list_regex_pattern_sets

Main.Wafv2.list_resources_for_web_acl

Main.Wafv2.list_rule_groups

Main.Wafv2.list_tags_for_resource

Main.Wafv2.list_web_acls

Main.Wafv2.put_logging_configuration

Main.Wafv2.put_permission_policy

Main.Wafv2.tag_resource

Main.Wafv2.untag_resource

Main.Wafv2.update_ipset

Main.Wafv2.update_regex_pattern_set

Main.Wafv2.update_rule_group

Main.Wafv2.update_web_acl

Documentation

Main.Wafv2.associate_web_acl — Method

associate_web_acl(resource_arn, web_aclarn)
associate_web_acl(resource_arn, web_aclarn, params::Dict{String,<:Any})

Associates a Web ACL with a regional application resource, to protect the resource. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. For AWS CloudFront, don't use this call. Instead, use your CloudFront distribution configuration. To associate a Web ACL, in the CloudFront call UpdateDistribution, set the web ACL ID to the Amazon Resource Name (ARN) of the Web ACL. For information, see UpdateDistribution.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource to associate with the web ACL. The ARN must be in one of the following formats: For an Application Load Balancer: arn:aws:elasticloadbalancing:region:account-id:loadbalancer/app/load-balancer-name/load-bala ncer-id For an API Gateway REST API: arn:aws:apigateway:region::/restapis/api-id/stages/stage-name For an AppSync GraphQL API: arn:aws:appsync:region:account-id:apis/GraphQLApiId

web_aclarn: The Amazon Resource Name (ARN) of the Web ACL that you want to associate with the resource.

Main.Wafv2.check_capacity — Method

check_capacity(rules, scope)
check_capacity(rules, scope, params::Dict{String,<:Any})

Returns the web ACL capacity unit (WCU) requirements for a specified scope and set of rules. You can use this to check the capacity requirements for the rules you want to use in a RuleGroup or WebACL. AWS WAF uses WCUs to calculate and control the operating resources that are used to run your rules, rule groups, and web ACLs. AWS WAF calculates capacity differently for each rule type, to reflect the relative cost of each rule. Simple rules that cost little to run use fewer WCUs than more complex rules that use more processing power. Rule group capacity is fixed at creation, which helps users plan their web ACL WCU usage when they use a rule group. The WCU limit for web ACLs is 1,500.

Arguments

rules: An array of Rule that you're configuring to use in a rule group or web ACL.

scope: Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

Main.Wafv2.create_ipset — Method

create_ipset(addresses, ipaddress_version, name, scope)
create_ipset(addresses, ipaddress_version, name, scope, params::Dict{String,<:Any})

Creates an IPSet, which you use to identify web requests that originate from specific IP addresses or ranges of IP addresses. For example, if you're receiving a lot of requests from a ranges of IP addresses, you can configure AWS WAF to block them using an IPSet that lists those IP addresses.

Arguments

addresses: Contains an array of strings that specify one or more IP addresses or blocks of IP addresses in Classless Inter-Domain Routing (CIDR) notation. AWS WAF supports all address ranges for IP versions IPv4 and IPv6. Examples: To configure AWS WAF to allow, block, or count requests that originated from the IP address 192.0.2.44, specify 192.0.2.44/32. To configure AWS WAF to allow, block, or count requests that originated from IP addresses from 192.0.2.0 to 192.0.2.255, specify 192.0.2.0/24. To configure AWS WAF to allow, block, or count requests that originated from the IP address 1111:0000:0000:0000:0000:0000:0000:0111, specify 1111:0000:0000:0000:0000:0000:0000:0111/128. To configure AWS WAF to allow, block, or count requests that originated from IP addresses 1111:0000:0000:0000:0000:0000:0000:0000 to 1111:0000:0000:0000:ffff:ffff:ffff:ffff, specify 1111:0000:0000:0000:0000:0000:0000:0000/64. For more information about CIDR notation, see the Wikipedia entry Classless Inter-Domain Routing.

ipaddress_version: Specify IPV4 or IPV6.

name: The name of the IP set. You cannot change the name of an IPSet after you create it.

scope: Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A description of the IP set that helps with identification.

"Tags": An array of key:value pairs to associate with the resource.

Main.Wafv2.create_regex_pattern_set — Method

create_regex_pattern_set(name, regular_expression_list, scope)
create_regex_pattern_set(name, regular_expression_list, scope, params::Dict{String,<:Any})

Creates a RegexPatternSet, which you reference in a RegexPatternSetReferenceStatement, to have AWS WAF inspect a web request component for the specified patterns.

Arguments

name: The name of the set. You cannot change the name after you create the set.

regular_expression_list: Array of regular expression strings.

scope: Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A description of the set that helps with identification.

"Tags": An array of key:value pairs to associate with the resource.

Main.Wafv2.create_rule_group — Method

create_rule_group(capacity, name, scope, visibility_config)
create_rule_group(capacity, name, scope, visibility_config, params::Dict{String,<:Any})

Creates a RuleGroup per the specifications provided. A rule group defines a collection of rules to inspect and control web requests that you can use in a WebACL. When you create a rule group, you define an immutable capacity limit. If you update a rule group, you must stay within the capacity. This allows others to reuse the rule group with confidence in its capacity requirements.

Arguments

capacity: The web ACL capacity units (WCUs) required for this rule group. When you create your own rule group, you define this, and you cannot change it after creation. When you add or modify the rules in a rule group, AWS WAF enforces this limit. You can check the capacity for a set of rules using CheckCapacity. AWS WAF uses WCUs to calculate and control the operating resources that are used to run your rules, rule groups, and web ACLs. AWS WAF calculates capacity differently for each rule type, to reflect the relative cost of each rule. Simple rules that cost little to run use fewer WCUs than more complex rules that use more processing power. Rule group capacity is fixed at creation, which helps users plan their web ACL WCU usage when they use a rule group. The WCU limit for web ACLs is 1,500.

name: The name of the rule group. You cannot change the name of a rule group after you create it.

scope: Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

visibility_config: Defines and enables Amazon CloudWatch metrics and web request sample collection.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A description of the rule group that helps with identification.

"Rules": The Rule statements used to identify the web requests that you want to allow, block, or count. Each rule includes one top-level statement that AWS WAF uses to identify matching web requests, and parameters that govern how AWS WAF handles them.

"Tags": An array of key:value pairs to associate with the resource.

Main.Wafv2.create_web_acl — Method

create_web_acl(default_action, name, scope, visibility_config)
create_web_acl(default_action, name, scope, visibility_config, params::Dict{String,<:Any})

Creates a WebACL per the specifications provided. A Web ACL defines a collection of rules to use to inspect and control web requests. Each rule has an action defined (allow, block, or count) for requests that match the statement of the rule. In the Web ACL, you assign a default action to take (allow, block) for any request that does not match any of the rules. The rules in a Web ACL can be a combination of the types Rule, RuleGroup, and managed rule group. You can associate a Web ACL with one or more AWS resources to protect. The resources can be Amazon CloudFront, an Amazon API Gateway REST API, an Application Load Balancer, or an AWS AppSync GraphQL API.

Arguments

default_action: The action to perform if none of the Rules contained in the WebACL match.

name: The name of the Web ACL. You cannot change the name of a Web ACL after you create it.

scope: Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

visibility_config: Defines and enables Amazon CloudWatch metrics and web request sample collection.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A description of the Web ACL that helps with identification.

"Rules": The Rule statements used to identify the web requests that you want to allow, block, or count. Each rule includes one top-level statement that AWS WAF uses to identify matching web requests, and parameters that govern how AWS WAF handles them.

"Tags": An array of key:value pairs to associate with the resource.

Main.Wafv2.delete_firewall_manager_rule_groups — Method

delete_firewall_manager_rule_groups(web_aclarn, web_acllock_token)
delete_firewall_manager_rule_groups(web_aclarn, web_acllock_token, params::Dict{String,<:Any})

Deletes all rule groups that are managed by AWS Firewall Manager for the specified web ACL. You can only use this if ManagedByFirewallManager is false in the specified WebACL.

Arguments

web_aclarn: The Amazon Resource Name (ARN) of the web ACL.

web_acllock_token: A token used for optimistic locking. AWS WAF returns a token to your get and list requests, to mark the state of the entity at the time of the request. To make changes to the entity associated with the token, you provide the token to operations like update and delete. AWS WAF uses the token to ensure that no changes have been made to the entity since you last retrieved it. If a change has been made, the update fails with a WAFOptimisticLockException. If this happens, perform another get, and use the new token returned by that operation.

Main.Wafv2.delete_ipset — Method

delete_ipset(id, lock_token, name, scope)
delete_ipset(id, lock_token, name, scope, params::Dict{String,<:Any})

Deletes the specified IPSet.

Arguments

id: A unique identifier for the set. This ID is returned in the responses to create and list commands. You provide it to operations like update and delete.

lock_token: A token used for optimistic locking. AWS WAF returns a token to your get and list requests, to mark the state of the entity at the time of the request. To make changes to the entity associated with the token, you provide the token to operations like update and delete. AWS WAF uses the token to ensure that no changes have been made to the entity since you last retrieved it. If a change has been made, the update fails with a WAFOptimisticLockException. If this happens, perform another get, and use the new token returned by that operation.

name: The name of the IP set. You cannot change the name of an IPSet after you create it.

scope: Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

Main.Wafv2.delete_logging_configuration — Method

delete_logging_configuration(resource_arn)
delete_logging_configuration(resource_arn, params::Dict{String,<:Any})

Deletes the LoggingConfiguration from the specified web ACL.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the web ACL from which you want to delete the LoggingConfiguration.

Main.Wafv2.delete_permission_policy — Method

delete_permission_policy(resource_arn)
delete_permission_policy(resource_arn, params::Dict{String,<:Any})

Permanently deletes an IAM policy from the specified rule group. You must be the owner of the rule group to perform this operation.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the rule group from which you want to delete the policy. You must be the owner of the rule group to perform this operation.

Main.Wafv2.delete_regex_pattern_set — Method

delete_regex_pattern_set(id, lock_token, name, scope)
delete_regex_pattern_set(id, lock_token, name, scope, params::Dict{String,<:Any})

Deletes the specified RegexPatternSet.

Arguments

id: A unique identifier for the set. This ID is returned in the responses to create and list commands. You provide it to operations like update and delete.

lock_token: A token used for optimistic locking. AWS WAF returns a token to your get and list requests, to mark the state of the entity at the time of the request. To make changes to the entity associated with the token, you provide the token to operations like update and delete. AWS WAF uses the token to ensure that no changes have been made to the entity since you last retrieved it. If a change has been made, the update fails with a WAFOptimisticLockException. If this happens, perform another get, and use the new token returned by that operation.

name: The name of the set. You cannot change the name after you create the set.

scope: Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

Main.Wafv2.delete_rule_group — Method

delete_rule_group(id, lock_token, name, scope)
delete_rule_group(id, lock_token, name, scope, params::Dict{String,<:Any})

Deletes the specified RuleGroup.

Arguments

id: A unique identifier for the rule group. This ID is returned in the responses to create and list commands. You provide it to operations like update and delete.

lock_token: A token used for optimistic locking. AWS WAF returns a token to your get and list requests, to mark the state of the entity at the time of the request. To make changes to the entity associated with the token, you provide the token to operations like update and delete. AWS WAF uses the token to ensure that no changes have been made to the entity since you last retrieved it. If a change has been made, the update fails with a WAFOptimisticLockException. If this happens, perform another get, and use the new token returned by that operation.

name: The name of the rule group. You cannot change the name of a rule group after you create it.

scope: Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

Main.Wafv2.delete_web_acl — Method

delete_web_acl(id, lock_token, name, scope)
delete_web_acl(id, lock_token, name, scope, params::Dict{String,<:Any})

Deletes the specified WebACL. You can only use this if ManagedByFirewallManager is false in the specified WebACL.

Arguments

id: The unique identifier for the Web ACL. This ID is returned in the responses to create and list commands. You provide it to operations like update and delete.

lock_token: A token used for optimistic locking. AWS WAF returns a token to your get and list requests, to mark the state of the entity at the time of the request. To make changes to the entity associated with the token, you provide the token to operations like update and delete. AWS WAF uses the token to ensure that no changes have been made to the entity since you last retrieved it. If a change has been made, the update fails with a WAFOptimisticLockException. If this happens, perform another get, and use the new token returned by that operation.

name: The name of the Web ACL. You cannot change the name of a Web ACL after you create it.

scope: Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

Main.Wafv2.describe_managed_rule_group — Method

describe_managed_rule_group(name, scope, vendor_name)
describe_managed_rule_group(name, scope, vendor_name, params::Dict{String,<:Any})

Provides high-level information for a managed rule group, including descriptions of the rules.

Arguments

name: The name of the managed rule group. You use this, along with the vendor name, to identify the rule group.

scope: Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

vendor_name: The name of the managed rule group vendor. You use this, along with the rule group name, to identify the rule group.

Main.Wafv2.disassociate_web_acl — Method

disassociate_web_acl(resource_arn)
disassociate_web_acl(resource_arn, params::Dict{String,<:Any})

Disassociates a Web ACL from a regional application resource. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. For AWS CloudFront, don't use this call. Instead, use your CloudFront distribution configuration. To disassociate a Web ACL, provide an empty web ACL ID in the CloudFront call UpdateDistribution. For information, see UpdateDistribution.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource to disassociate from the web ACL. The ARN must be in one of the following formats: For an Application Load Balancer: arn:aws:elasticloadbalancing:region:account-id:loadbalancer/app/load-balancer-name/load-bala ncer-id For an API Gateway REST API: arn:aws:apigateway:region::/restapis/api-id/stages/stage-name For an AppSync GraphQL API: arn:aws:appsync:region:account-id:apis/GraphQLApiId

Main.Wafv2.get_ipset — Method

get_ipset(id, name, scope)
get_ipset(id, name, scope, params::Dict{String,<:Any})

Retrieves the specified IPSet.

Arguments

id: A unique identifier for the set. This ID is returned in the responses to create and list commands. You provide it to operations like update and delete.

name: The name of the IP set. You cannot change the name of an IPSet after you create it.

scope: Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

Main.Wafv2.get_logging_configuration — Method

get_logging_configuration(resource_arn)
get_logging_configuration(resource_arn, params::Dict{String,<:Any})

Returns the LoggingConfiguration for the specified web ACL.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the web ACL for which you want to get the LoggingConfiguration.

Main.Wafv2.get_permission_policy — Method

get_permission_policy(resource_arn)
get_permission_policy(resource_arn, params::Dict{String,<:Any})

Returns the IAM policy that is attached to the specified rule group. You must be the owner of the rule group to perform this operation.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the rule group for which you want to get the policy.

Main.Wafv2.get_rate_based_statement_managed_keys — Method

get_rate_based_statement_managed_keys(rule_name, scope, web_aclid, web_aclname)
get_rate_based_statement_managed_keys(rule_name, scope, web_aclid, web_aclname, params::Dict{String,<:Any})

Retrieves the keys that are currently blocked by a rate-based rule. The maximum number of managed keys that can be blocked for a single rate-based rule is 10,000. If more than 10,000 addresses exceed the rate limit, those with the highest rates are blocked.

Arguments

rule_name: The name of the rate-based rule to get the keys for.

scope: Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

web_aclid: The unique identifier for the Web ACL. This ID is returned in the responses to create and list commands. You provide it to operations like update and delete.

web_aclname: The name of the Web ACL. You cannot change the name of a Web ACL after you create it.

Main.Wafv2.get_regex_pattern_set — Method

get_regex_pattern_set(id, name, scope)
get_regex_pattern_set(id, name, scope, params::Dict{String,<:Any})

Retrieves the specified RegexPatternSet.

Arguments

id: A unique identifier for the set. This ID is returned in the responses to create and list commands. You provide it to operations like update and delete.

name: The name of the set. You cannot change the name after you create the set.

scope: Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

Main.Wafv2.get_rule_group — Method

get_rule_group(id, name, scope)
get_rule_group(id, name, scope, params::Dict{String,<:Any})

Retrieves the specified RuleGroup.

Arguments

id: A unique identifier for the rule group. This ID is returned in the responses to create and list commands. You provide it to operations like update and delete.

name: The name of the rule group. You cannot change the name of a rule group after you create it.

scope: Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

Main.Wafv2.get_sampled_requests — Method

get_sampled_requests(max_items, rule_metric_name, scope, time_window, web_acl_arn)
get_sampled_requests(max_items, rule_metric_name, scope, time_window, web_acl_arn, params::Dict{String,<:Any})

Gets detailed information about a specified number of requests–a sample–that AWS WAF randomly selects from among the first 5,000 requests that your AWS resource received during a time range that you choose. You can specify a sample size of up to 500 requests, and you can specify any time range in the previous three hours. GetSampledRequests returns a time range, which is usually the time range that you specified. However, if your resource (such as a CloudFront distribution) received 5,000 requests before the specified time range elapsed, GetSampledRequests returns an updated time range. This new time range indicates the actual period during which AWS WAF selected the requests in the sample.

Arguments

max_items: The number of requests that you want AWS WAF to return from among the first 5,000 requests that your AWS resource received during the time range. If your resource received fewer requests than the value of MaxItems, GetSampledRequests returns information about all of them.

rule_metric_name: The metric name assigned to the Rule or RuleGroup for which you want a sample of requests.

scope: Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

time_window: The start date and time and the end date and time of the range for which you want GetSampledRequests to return a sample of requests. You must specify the times in Coordinated Universal Time (UTC) format. UTC format includes the special designator, Z. For example, "2016-09-27T14:50Z". You can specify any time range in the previous three hours. If you specify a start time that's earlier than three hours ago, AWS WAF sets it to three hours ago.

web_acl_arn: The Amazon resource name (ARN) of the WebACL for which you want a sample of requests.

Main.Wafv2.get_web_acl — Method

get_web_acl(id, name, scope)
get_web_acl(id, name, scope, params::Dict{String,<:Any})

Retrieves the specified WebACL.

Arguments

id: The unique identifier for the Web ACL. This ID is returned in the responses to create and list commands. You provide it to operations like update and delete.

name: The name of the Web ACL. You cannot change the name of a Web ACL after you create it.

scope: Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

Main.Wafv2.get_web_aclfor_resource — Method

get_web_aclfor_resource(resource_arn)
get_web_aclfor_resource(resource_arn, params::Dict{String,<:Any})

Retrieves the WebACL for the specified resource.

Arguments

resource_arn: The ARN (Amazon Resource Name) of the resource.

Main.Wafv2.list_available_managed_rule_groups — Method

list_available_managed_rule_groups(scope)
list_available_managed_rule_groups(scope, params::Dict{String,<:Any})

Retrieves an array of managed rule groups that are available for you to use. This list includes all AWS Managed Rules rule groups and the AWS Marketplace managed rule groups that you're subscribed to.

Arguments

scope: Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Limit": The maximum number of objects that you want AWS WAF to return for this request. If more objects are available, in the response, AWS WAF provides a NextMarker value that you can use in a subsequent call to get the next batch of objects.

"NextMarker": When you request a list of objects with a Limit setting, if the number of objects that are still available for retrieval exceeds the limit, AWS WAF returns a NextMarker value in the response. To retrieve the next batch of objects, provide the marker from the prior call in your next request.

Main.Wafv2.list_ipsets — Method

list_ipsets(scope)
list_ipsets(scope, params::Dict{String,<:Any})

Retrieves an array of IPSetSummary objects for the IP sets that you manage.

Arguments

scope: Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Limit": The maximum number of objects that you want AWS WAF to return for this request. If more objects are available, in the response, AWS WAF provides a NextMarker value that you can use in a subsequent call to get the next batch of objects.

"NextMarker": When you request a list of objects with a Limit setting, if the number of objects that are still available for retrieval exceeds the limit, AWS WAF returns a NextMarker value in the response. To retrieve the next batch of objects, provide the marker from the prior call in your next request.

Main.Wafv2.list_logging_configurations — Method

list_logging_configurations()
list_logging_configurations(params::Dict{String,<:Any})

Retrieves an array of your LoggingConfiguration objects.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Limit": The maximum number of objects that you want AWS WAF to return for this request. If more objects are available, in the response, AWS WAF provides a NextMarker value that you can use in a subsequent call to get the next batch of objects.

"NextMarker": When you request a list of objects with a Limit setting, if the number of objects that are still available for retrieval exceeds the limit, AWS WAF returns a NextMarker value in the response. To retrieve the next batch of objects, provide the marker from the prior call in your next request.

"Scope": Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

Main.Wafv2.list_regex_pattern_sets — Method

list_regex_pattern_sets(scope)
list_regex_pattern_sets(scope, params::Dict{String,<:Any})

Retrieves an array of RegexPatternSetSummary objects for the regex pattern sets that you manage.

Arguments

scope: Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Limit": The maximum number of objects that you want AWS WAF to return for this request. If more objects are available, in the response, AWS WAF provides a NextMarker value that you can use in a subsequent call to get the next batch of objects.

"NextMarker": When you request a list of objects with a Limit setting, if the number of objects that are still available for retrieval exceeds the limit, AWS WAF returns a NextMarker value in the response. To retrieve the next batch of objects, provide the marker from the prior call in your next request.

Main.Wafv2.list_resources_for_web_acl — Method

list_resources_for_web_acl(web_aclarn)
list_resources_for_web_acl(web_aclarn, params::Dict{String,<:Any})

Retrieves an array of the Amazon Resource Names (ARNs) for the regional resources that are associated with the specified web ACL. If you want the list of AWS CloudFront resources, use the AWS CloudFront call ListDistributionsByWebACLId.

Arguments

web_aclarn: The Amazon Resource Name (ARN) of the Web ACL.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ResourceType": Used for web ACLs that are scoped for regional applications. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API.

Main.Wafv2.list_rule_groups — Method

list_rule_groups(scope)
list_rule_groups(scope, params::Dict{String,<:Any})

Retrieves an array of RuleGroupSummary objects for the rule groups that you manage.

Arguments

scope: Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Limit": The maximum number of objects that you want AWS WAF to return for this request. If more objects are available, in the response, AWS WAF provides a NextMarker value that you can use in a subsequent call to get the next batch of objects.

"NextMarker": When you request a list of objects with a Limit setting, if the number of objects that are still available for retrieval exceeds the limit, AWS WAF returns a NextMarker value in the response. To retrieve the next batch of objects, provide the marker from the prior call in your next request.

Main.Wafv2.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Retrieves the TagInfoForResource for the specified resource. Tags are key:value pairs that you can use to categorize and manage your resources, for purposes like billing. For example, you might set the tag key to "customer" and the value to the customer name or ID. You can specify one or more tags to add to each AWS resource, up to 50 tags for a resource. You can tag the AWS resources that you manage through AWS WAF: web ACLs, rule groups, IP sets, and regex pattern sets. You can't manage or view tags through the AWS WAF console.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Limit": The maximum number of objects that you want AWS WAF to return for this request. If more objects are available, in the response, AWS WAF provides a NextMarker value that you can use in a subsequent call to get the next batch of objects.

"NextMarker": When you request a list of objects with a Limit setting, if the number of objects that are still available for retrieval exceeds the limit, AWS WAF returns a NextMarker value in the response. To retrieve the next batch of objects, provide the marker from the prior call in your next request.

Main.Wafv2.list_web_acls — Method

list_web_acls(scope)
list_web_acls(scope, params::Dict{String,<:Any})

Retrieves an array of WebACLSummary objects for the web ACLs that you manage.

Arguments

scope: Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Limit": The maximum number of objects that you want AWS WAF to return for this request. If more objects are available, in the response, AWS WAF provides a NextMarker value that you can use in a subsequent call to get the next batch of objects.

"NextMarker": When you request a list of objects with a Limit setting, if the number of objects that are still available for retrieval exceeds the limit, AWS WAF returns a NextMarker value in the response. To retrieve the next batch of objects, provide the marker from the prior call in your next request.

Main.Wafv2.put_logging_configuration — Method

put_logging_configuration(logging_configuration)
put_logging_configuration(logging_configuration, params::Dict{String,<:Any})

Enables the specified LoggingConfiguration, to start logging from a web ACL, according to the configuration provided. You can access information about all traffic that AWS WAF inspects using the following steps: Create an Amazon Kinesis Data Firehose. Create the data firehose with a PUT source and in the Region that you are operating. If you are capturing logs for Amazon CloudFront, always create the firehose in US East (N. Virginia). Give the data firehose a name that starts with the prefix aws-waf-logs-. For example, aws-waf-logs-us-east-2-analytics. Do not create the data firehose using a Kinesis stream as your source. Associate that firehose to your web ACL using a PutLoggingConfiguration request. When you successfully enable logging using a PutLoggingConfiguration request, AWS WAF will create a service linked role with the necessary permissions to write logs to the Amazon Kinesis Data Firehose. For more information, see Logging Web ACL Traffic Information in the AWS WAF Developer Guide.

Arguments

logging_configuration:

Main.Wafv2.put_permission_policy — Method

put_permission_policy(policy, resource_arn)
put_permission_policy(policy, resource_arn, params::Dict{String,<:Any})

Attaches an IAM policy to the specified resource. Use this to share a rule group across accounts. You must be the owner of the rule group to perform this operation. This action is subject to the following restrictions: You can attach only one policy with each PutPermissionPolicy request. The ARN in the request must be a valid WAF RuleGroup ARN and the rule group must exist in the same region. The user making the request must be the owner of the rule group.

Arguments

policy: The policy to attach to the specified rule group. The policy specifications must conform to the following: The policy must be composed using IAM Policy version 2012-10-17 or version 2015-01-01. The policy must include specifications for Effect, Action, and Principal. Effect must specify Allow. Action must specify wafv2:CreateWebACL, wafv2:UpdateWebACL, and wafv2:PutFirewallManagerRuleGroups. AWS WAF rejects any extra actions or wildcard actions in the policy. The policy must not include a Resource parameter. For more information, see IAM Policies.

resource_arn: The Amazon Resource Name (ARN) of the RuleGroup to which you want to attach the policy.

Main.Wafv2.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Associates tags with the specified AWS resource. Tags are key:value pairs that you can use to categorize and manage your resources, for purposes like billing. For example, you might set the tag key to "customer" and the value to the customer name or ID. You can specify one or more tags to add to each AWS resource, up to 50 tags for a resource. You can tag the AWS resources that you manage through AWS WAF: web ACLs, rule groups, IP sets, and regex pattern sets. You can't manage or view tags through the AWS WAF console.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

tags: An array of key:value pairs to associate with the resource.

Main.Wafv2.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Disassociates tags from an AWS resource. Tags are key:value pairs that you can associate with AWS resources. For example, the tag key might be "customer" and the tag value might be "companyA." You can specify one or more tags to add to each container. You can add up to 50 tags to each AWS resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

tag_keys: An array of keys identifying the tags to disassociate from the resource.

Main.Wafv2.update_ipset — Method

update_ipset(addresses, id, lock_token, name, scope)
update_ipset(addresses, id, lock_token, name, scope, params::Dict{String,<:Any})

Updates the specified IPSet.

Arguments

addresses: Contains an array of strings that specify one or more IP addresses or blocks of IP addresses in Classless Inter-Domain Routing (CIDR) notation. AWS WAF supports all address ranges for IP versions IPv4 and IPv6. Examples: To configure AWS WAF to allow, block, or count requests that originated from the IP address 192.0.2.44, specify 192.0.2.44/32. To configure AWS WAF to allow, block, or count requests that originated from IP addresses from 192.0.2.0 to 192.0.2.255, specify 192.0.2.0/24. To configure AWS WAF to allow, block, or count requests that originated from the IP address 1111:0000:0000:0000:0000:0000:0000:0111, specify 1111:0000:0000:0000:0000:0000:0000:0111/128. To configure AWS WAF to allow, block, or count requests that originated from IP addresses 1111:0000:0000:0000:0000:0000:0000:0000 to 1111:0000:0000:0000:ffff:ffff:ffff:ffff, specify 1111:0000:0000:0000:0000:0000:0000:0000/64. For more information about CIDR notation, see the Wikipedia entry Classless Inter-Domain Routing.

id: A unique identifier for the set. This ID is returned in the responses to create and list commands. You provide it to operations like update and delete.

lock_token: A token used for optimistic locking. AWS WAF returns a token to your get and list requests, to mark the state of the entity at the time of the request. To make changes to the entity associated with the token, you provide the token to operations like update and delete. AWS WAF uses the token to ensure that no changes have been made to the entity since you last retrieved it. If a change has been made, the update fails with a WAFOptimisticLockException. If this happens, perform another get, and use the new token returned by that operation.

name: The name of the IP set. You cannot change the name of an IPSet after you create it.

scope: Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A description of the IP set that helps with identification.

Main.Wafv2.update_regex_pattern_set — Method

update_regex_pattern_set(id, lock_token, name, regular_expression_list, scope)
update_regex_pattern_set(id, lock_token, name, regular_expression_list, scope, params::Dict{String,<:Any})

Updates the specified RegexPatternSet.

Arguments

id: A unique identifier for the set. This ID is returned in the responses to create and list commands. You provide it to operations like update and delete.

lock_token: A token used for optimistic locking. AWS WAF returns a token to your get and list requests, to mark the state of the entity at the time of the request. To make changes to the entity associated with the token, you provide the token to operations like update and delete. AWS WAF uses the token to ensure that no changes have been made to the entity since you last retrieved it. If a change has been made, the update fails with a WAFOptimisticLockException. If this happens, perform another get, and use the new token returned by that operation.

name: The name of the set. You cannot change the name after you create the set.

regular_expression_list:

scope: Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A description of the set that helps with identification.

Main.Wafv2.update_rule_group — Method

update_rule_group(id, lock_token, name, scope, visibility_config)
update_rule_group(id, lock_token, name, scope, visibility_config, params::Dict{String,<:Any})

Updates the specified RuleGroup. A rule group defines a collection of rules to inspect and control web requests that you can use in a WebACL. When you create a rule group, you define an immutable capacity limit. If you update a rule group, you must stay within the capacity. This allows others to reuse the rule group with confidence in its capacity requirements.

Arguments

id: A unique identifier for the rule group. This ID is returned in the responses to create and list commands. You provide it to operations like update and delete.

lock_token: A token used for optimistic locking. AWS WAF returns a token to your get and list requests, to mark the state of the entity at the time of the request. To make changes to the entity associated with the token, you provide the token to operations like update and delete. AWS WAF uses the token to ensure that no changes have been made to the entity since you last retrieved it. If a change has been made, the update fails with a WAFOptimisticLockException. If this happens, perform another get, and use the new token returned by that operation.

name: The name of the rule group. You cannot change the name of a rule group after you create it.

scope: Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

visibility_config: Defines and enables Amazon CloudWatch metrics and web request sample collection.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A description of the rule group that helps with identification.

"Rules": The Rule statements used to identify the web requests that you want to allow, block, or count. Each rule includes one top-level statement that AWS WAF uses to identify matching web requests, and parameters that govern how AWS WAF handles them.

Main.Wafv2.update_web_acl — Method

update_web_acl(default_action, id, lock_token, name, scope, visibility_config)
update_web_acl(default_action, id, lock_token, name, scope, visibility_config, params::Dict{String,<:Any})

Updates the specified WebACL. A Web ACL defines a collection of rules to use to inspect and control web requests. Each rule has an action defined (allow, block, or count) for requests that match the statement of the rule. In the Web ACL, you assign a default action to take (allow, block) for any request that does not match any of the rules. The rules in a Web ACL can be a combination of the types Rule, RuleGroup, and managed rule group. You can associate a Web ACL with one or more AWS resources to protect. The resources can be Amazon CloudFront, an Amazon API Gateway REST API, an Application Load Balancer, or an AWS AppSync GraphQL API.

Arguments

default_action: The action to perform if none of the Rules contained in the WebACL match.

id: The unique identifier for the Web ACL. This ID is returned in the responses to create and list commands. You provide it to operations like update and delete.

lock_token: A token used for optimistic locking. AWS WAF returns a token to your get and list requests, to mark the state of the entity at the time of the request. To make changes to the entity associated with the token, you provide the token to operations like update and delete. AWS WAF uses the token to ensure that no changes have been made to the entity since you last retrieved it. If a change has been made, the update fails with a WAFOptimisticLockException. If this happens, perform another get, and use the new token returned by that operation.

name: The name of the Web ACL. You cannot change the name of a Web ACL after you create it.

scope: Specifies whether this is for an AWS CloudFront distribution or for a regional application. A regional application can be an Application Load Balancer (ALB), an API Gateway REST API, or an AppSync GraphQL API. To work with CloudFront, you must also specify the Region US East (N. Virginia) as follows: CLI - Specify the Region when you use the CloudFront scope: –scope=CLOUDFRONT –region=us-east-1. API and SDKs - For all calls, use the Region endpoint us-east-1.

visibility_config: Defines and enables Amazon CloudWatch metrics and web request sample collection.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A description of the Web ACL that helps with identification.

"Rules": The Rule statements used to identify the web requests that you want to allow, block, or count. Each rule includes one top-level statement that AWS WAF uses to identify matching web requests, and parameters that govern how AWS WAF handles them.