Workdocs

This page documents function available when using the Workdocs module, created with @service Workdocs.

Index

Main.Workdocs.abort_document_version_upload

Main.Workdocs.activate_user

Main.Workdocs.add_resource_permissions

Main.Workdocs.create_comment

Main.Workdocs.create_custom_metadata

Main.Workdocs.create_folder

Main.Workdocs.create_labels

Main.Workdocs.create_notification_subscription

Main.Workdocs.create_user

Main.Workdocs.deactivate_user

Main.Workdocs.delete_comment

Main.Workdocs.delete_custom_metadata

Main.Workdocs.delete_document

Main.Workdocs.delete_folder

Main.Workdocs.delete_folder_contents

Main.Workdocs.delete_labels

Main.Workdocs.delete_notification_subscription

Main.Workdocs.delete_user

Main.Workdocs.describe_activities

Main.Workdocs.describe_comments

Main.Workdocs.describe_document_versions

Main.Workdocs.describe_folder_contents

Main.Workdocs.describe_groups

Main.Workdocs.describe_notification_subscriptions

Main.Workdocs.describe_resource_permissions

Main.Workdocs.describe_root_folders

Main.Workdocs.describe_users

Main.Workdocs.get_current_user

Main.Workdocs.get_document

Main.Workdocs.get_document_path

Main.Workdocs.get_document_version

Main.Workdocs.get_folder

Main.Workdocs.get_folder_path

Main.Workdocs.get_resources

Main.Workdocs.initiate_document_version_upload

Main.Workdocs.remove_all_resource_permissions

Main.Workdocs.remove_resource_permission

Main.Workdocs.update_document

Main.Workdocs.update_document_version

Main.Workdocs.update_folder

Main.Workdocs.update_user

Documentation

Main.Workdocs.abort_document_version_upload — Method

abort_document_version_upload(document_id, version_id)
abort_document_version_upload(document_id, version_id, params::Dict{String,<:Any})

Aborts the upload of the specified document version that was previously initiated by InitiateDocumentVersionUpload. The client should make this call only when it no longer intends to upload the document version, or fails to do so.

Arguments

document_id: The ID of the document.

version_id: The ID of the version.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

Main.Workdocs.activate_user — Method

activate_user(user_id)
activate_user(user_id, params::Dict{String,<:Any})

Activates the specified user. Only active users can access Amazon WorkDocs.

Arguments

user_id: The ID of the user.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

Main.Workdocs.add_resource_permissions — Method

add_resource_permissions(principals, resource_id)
add_resource_permissions(principals, resource_id, params::Dict{String,<:Any})

Creates a set of permissions for the specified folder or document. The resource permissions are overwritten if the principals already have different permissions.

Arguments

principals: The users, groups, or organization being granted permission.

resource_id: The ID of the resource.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"NotificationOptions": The notification options.

Main.Workdocs.create_comment — Method

create_comment(document_id, text, version_id)
create_comment(document_id, text, version_id, params::Dict{String,<:Any})

Adds a new comment to the specified document version.

Arguments

document_id: The ID of the document.

text: The text of the comment.

version_id: The ID of the document version.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"NotifyCollaborators": Set this parameter to TRUE to send an email out to the document collaborators after the comment is created.

"ParentId": The ID of the parent comment.

"ThreadId": The ID of the root comment in the thread.

"Visibility": The visibility of the comment. Options are either PRIVATE, where the comment is visible only to the comment author and document owner and co-owners, or PUBLIC, where the comment is visible to document owners, co-owners, and contributors.

Main.Workdocs.create_custom_metadata — Method

create_custom_metadata(custom_metadata, resource_id)
create_custom_metadata(custom_metadata, resource_id, params::Dict{String,<:Any})

Adds one or more custom properties to the specified resource (a folder, document, or version).

Arguments

custom_metadata: Custom metadata in the form of name-value pairs.

resource_id: The ID of the resource.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"versionid": The ID of the version, if the custom metadata is being added to a document version.

Main.Workdocs.create_folder — Method

create_folder(parent_folder_id)
create_folder(parent_folder_id, params::Dict{String,<:Any})

Creates a folder with the specified name and parent folder.

Arguments

parent_folder_id: The ID of the parent folder.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"Name": The name of the new folder.

Main.Workdocs.create_labels — Method

create_labels(labels, resource_id)
create_labels(labels, resource_id, params::Dict{String,<:Any})

Adds the specified list of labels to the given resource (a document or folder)

Arguments

labels: List of labels to add to the resource.

resource_id: The ID of the resource.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

Main.Workdocs.create_notification_subscription — Method

create_notification_subscription(endpoint, organization_id, protocol, subscription_type)
create_notification_subscription(endpoint, organization_id, protocol, subscription_type, params::Dict{String,<:Any})

Configure Amazon WorkDocs to use Amazon SNS notifications. The endpoint receives a confirmation message, and must confirm the subscription. For more information, see Subscribe to Notifications in the Amazon WorkDocs Developer Guide.

Arguments

endpoint: The endpoint to receive the notifications. If the protocol is HTTPS, the endpoint is a URL that begins with https.

organization_id: The ID of the organization.

protocol: The protocol to use. The supported value is https, which delivers JSON-encoded messages using HTTPS POST.

subscription_type: The notification type.

Main.Workdocs.create_user — Method

create_user(given_name, password, surname, username)
create_user(given_name, password, surname, username, params::Dict{String,<:Any})

Creates a user in a Simple AD or Microsoft AD directory. The status of a newly created user is "ACTIVE". New users can access Amazon WorkDocs.

Arguments

given_name: The given name of the user.

password: The password of the user.

surname: The surname of the user.

username: The login name of the user.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"EmailAddress": The email address of the user.

"OrganizationId": The ID of the organization.

"StorageRule": The amount of storage for the user.

"TimeZoneId": The time zone ID of the user.

Main.Workdocs.deactivate_user — Method

deactivate_user(user_id)
deactivate_user(user_id, params::Dict{String,<:Any})

Deactivates the specified user, which revokes the user's access to Amazon WorkDocs.

Arguments

user_id: The ID of the user.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

Main.Workdocs.delete_comment — Method

delete_comment(comment_id, document_id, version_id)
delete_comment(comment_id, document_id, version_id, params::Dict{String,<:Any})

Deletes the specified comment from the document version.

Arguments

comment_id: The ID of the comment.

document_id: The ID of the document.

version_id: The ID of the document version.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

Main.Workdocs.delete_custom_metadata — Method

delete_custom_metadata(resource_id)
delete_custom_metadata(resource_id, params::Dict{String,<:Any})

Deletes custom metadata from the specified resource.

Arguments

resource_id: The ID of the resource, either a document or folder.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"deleteAll": Flag to indicate removal of all custom metadata properties from the specified resource.

"keys": List of properties to remove.

"versionId": The ID of the version, if the custom metadata is being deleted from a document version.

Main.Workdocs.delete_document — Method

delete_document(document_id)
delete_document(document_id, params::Dict{String,<:Any})

Permanently deletes the specified document and its associated metadata.

Arguments

document_id: The ID of the document.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

Main.Workdocs.delete_folder — Method

delete_folder(folder_id)
delete_folder(folder_id, params::Dict{String,<:Any})

Permanently deletes the specified folder and its contents.

Arguments

folder_id: The ID of the folder.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

Main.Workdocs.delete_folder_contents — Method

delete_folder_contents(folder_id)
delete_folder_contents(folder_id, params::Dict{String,<:Any})

Deletes the contents of the specified folder.

Arguments

folder_id: The ID of the folder.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

Main.Workdocs.delete_labels — Method

delete_labels(resource_id)
delete_labels(resource_id, params::Dict{String,<:Any})

Deletes the specified list of labels from a resource.

Arguments

resource_id: The ID of the resource.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"deleteAll": Flag to request removal of all labels from the specified resource.

"labels": List of labels to delete from the resource.

Main.Workdocs.delete_notification_subscription — Method

delete_notification_subscription(organization_id, subscription_id)
delete_notification_subscription(organization_id, subscription_id, params::Dict{String,<:Any})

Deletes the specified subscription from the specified organization.

Arguments

organization_id: The ID of the organization.

subscription_id: The ID of the subscription.

Main.Workdocs.delete_user — Method

delete_user(user_id)
delete_user(user_id, params::Dict{String,<:Any})

Deletes the specified user from a Simple AD or Microsoft AD directory.

Arguments

user_id: The ID of the user.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Do not set this field when using administrative API actions, as in accessing the API using AWS credentials.

Main.Workdocs.describe_activities — Method

describe_activities()
describe_activities(params::Dict{String,<:Any})

Describes the user activities in a specified time period.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"activityTypes": Specifies which activity types to include in the response. If this field is left empty, all activity types are returned.

"endTime": The timestamp that determines the end time of the activities. The response includes the activities performed before the specified timestamp.

"includeIndirectActivities": Includes indirect activities. An indirect activity results from a direct activity performed on a parent resource. For example, sharing a parent folder (the direct activity) shares all of the subfolders and documents within the parent folder (the indirect activity).

"limit": The maximum number of items to return.

"marker": The marker for the next set of results.

"organizationId": The ID of the organization. This is a mandatory parameter when using administrative API (SigV4) requests.

"resourceId": The document or folder ID for which to describe activity types.

"startTime": The timestamp that determines the starting time of the activities. The response includes the activities performed after the specified timestamp.

"userId": The ID of the user who performed the action. The response includes activities pertaining to this user. This is an optional parameter and is only applicable for administrative API (SigV4) requests.

Main.Workdocs.describe_comments — Method

describe_comments(document_id, version_id)
describe_comments(document_id, version_id, params::Dict{String,<:Any})

List all the comments for the specified document version.

Arguments

document_id: The ID of the document.

version_id: The ID of the document version.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"limit": The maximum number of items to return.

"marker": The marker for the next set of results. This marker was received from a previous call.

Main.Workdocs.describe_document_versions — Method

describe_document_versions(document_id)
describe_document_versions(document_id, params::Dict{String,<:Any})

Retrieves the document versions for the specified document. By default, only active versions are returned.

Arguments

document_id: The ID of the document.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"fields": Specify "SOURCE" to include initialized versions and a URL for the source document.

"include": A comma-separated list of values. Specify "INITIALIZED" to include incomplete versions.

"limit": The maximum number of versions to return with this call.

"marker": The marker for the next set of results. (You received this marker from a previous call.)

Main.Workdocs.describe_folder_contents — Method

describe_folder_contents(folder_id)
describe_folder_contents(folder_id, params::Dict{String,<:Any})

Describes the contents of the specified folder, including its documents and subfolders. By default, Amazon WorkDocs returns the first 100 active document and folder metadata items. If there are more results, the response includes a marker that you can use to request the next set of results. You can also request initialized documents.

Arguments

folder_id: The ID of the folder.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"include": The contents to include. Specify "INITIALIZED" to include initialized documents.

"limit": The maximum number of items to return with this call.

"marker": The marker for the next set of results. This marker was received from a previous call.

"order": The order for the contents of the folder.

"sort": The sorting criteria.

"type": The type of items.

Main.Workdocs.describe_groups — Method

describe_groups(search_query)
describe_groups(search_query, params::Dict{String,<:Any})

Describes the groups specified by the query. Groups are defined by the underlying Active Directory.

Arguments

search_query: A query to describe groups by group name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"limit": The maximum number of items to return with this call.

"marker": The marker for the next set of results. (You received this marker from a previous call.)

"organizationId": The ID of the organization.

Main.Workdocs.describe_notification_subscriptions — Method

describe_notification_subscriptions(organization_id)
describe_notification_subscriptions(organization_id, params::Dict{String,<:Any})

Lists the specified notification subscriptions.

Arguments

organization_id: The ID of the organization.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"limit": The maximum number of items to return with this call.

"marker": The marker for the next set of results. (You received this marker from a previous call.)

Main.Workdocs.describe_resource_permissions — Method

describe_resource_permissions(resource_id)
describe_resource_permissions(resource_id, params::Dict{String,<:Any})

Describes the permissions of a specified resource.

Arguments

resource_id: The ID of the resource.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"limit": The maximum number of items to return with this call.

"marker": The marker for the next set of results. (You received this marker from a previous call)

"principalId": The ID of the principal to filter permissions by.

Main.Workdocs.describe_root_folders — Method

describe_root_folders(authentication)
describe_root_folders(authentication, params::Dict{String,<:Any})

Describes the current user's special folders; the RootFolder and the RecycleBin. RootFolder is the root of user's files and folders and RecycleBin is the root of recycled items. This is not a valid action for SigV4 (administrative API) clients. This action requires an authentication token. To get an authentication token, register an application with Amazon WorkDocs. For more information, see Authentication and Access Control for User Applications in the Amazon WorkDocs Developer Guide.

Arguments

authentication: Amazon WorkDocs authentication token.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"limit": The maximum number of items to return.

"marker": The marker for the next set of results. (You received this marker from a previous call.)

Main.Workdocs.describe_users — Method

describe_users()
describe_users(params::Dict{String,<:Any})

Describes the specified users. You can describe all users or filter the results (for example, by status or organization). By default, Amazon WorkDocs returns the first 24 active or pending users. If there are more results, the response includes a marker that you can use to request the next set of results.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"fields": A comma-separated list of values. Specify "STORAGE_METADATA" to include the user storage quota and utilization information.

"include": The state of the users. Specify "ALL" to include inactive users.

"limit": The maximum number of items to return.

"marker": The marker for the next set of results. (You received this marker from a previous call.)

"order": The order for the results.

"organizationId": The ID of the organization.

"query": A query to filter users by user name.

"sort": The sorting criteria.

"userIds": The IDs of the users.

Main.Workdocs.get_current_user — Method

get_current_user(authentication)
get_current_user(authentication, params::Dict{String,<:Any})

Retrieves details of the current user for whom the authentication token was generated. This is not a valid action for SigV4 (administrative API) clients. This action requires an authentication token. To get an authentication token, register an application with Amazon WorkDocs. For more information, see Authentication and Access Control for User Applications in the Amazon WorkDocs Developer Guide.

Arguments

authentication: Amazon WorkDocs authentication token.

Main.Workdocs.get_document — Method

get_document(document_id)
get_document(document_id, params::Dict{String,<:Any})

Retrieves details of a document.

Arguments

document_id: The ID of the document.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"includeCustomMetadata": Set this to TRUE to include custom metadata in the response.

Main.Workdocs.get_document_path — Method

get_document_path(document_id)
get_document_path(document_id, params::Dict{String,<:Any})

Retrieves the path information (the hierarchy from the root folder) for the requested document. By default, Amazon WorkDocs returns a maximum of 100 levels upwards from the requested document and only includes the IDs of the parent folders in the path. You can limit the maximum number of levels. You can also request the names of the parent folders.

Arguments

document_id: The ID of the document.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"fields": A comma-separated list of values. Specify NAME to include the names of the parent folders.

"limit": The maximum number of levels in the hierarchy to return.

"marker": This value is not supported.

Main.Workdocs.get_document_version — Method

get_document_version(document_id, version_id)
get_document_version(document_id, version_id, params::Dict{String,<:Any})

Retrieves version metadata for the specified document.

Arguments

document_id: The ID of the document.

version_id: The version ID of the document.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"fields": A comma-separated list of values. Specify "SOURCE" to include a URL for the source document.

"includeCustomMetadata": Set this to TRUE to include custom metadata in the response.

Main.Workdocs.get_folder — Method

get_folder(folder_id)
get_folder(folder_id, params::Dict{String,<:Any})

Retrieves the metadata of the specified folder.

Arguments

folder_id: The ID of the folder.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"includeCustomMetadata": Set to TRUE to include custom metadata in the response.

Main.Workdocs.get_folder_path — Method

get_folder_path(folder_id)
get_folder_path(folder_id, params::Dict{String,<:Any})

Retrieves the path information (the hierarchy from the root folder) for the specified folder. By default, Amazon WorkDocs returns a maximum of 100 levels upwards from the requested folder and only includes the IDs of the parent folders in the path. You can limit the maximum number of levels. You can also request the parent folder names.

Arguments

folder_id: The ID of the folder.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"fields": A comma-separated list of values. Specify "NAME" to include the names of the parent folders.

"limit": The maximum number of levels in the hierarchy to return.

"marker": This value is not supported.

Main.Workdocs.get_resources — Method

get_resources()
get_resources(params::Dict{String,<:Any})

Retrieves a collection of resources, including folders and documents. The only CollectionType supported is SHAREDWITHME.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": The Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"collectionType": The collection type.

"limit": The maximum number of resources to return.

"marker": The marker for the next set of results. This marker was received from a previous call.

"userId": The user ID for the resource collection. This is a required field for accessing the API operation using IAM credentials.

Main.Workdocs.initiate_document_version_upload — Method

initiate_document_version_upload(parent_folder_id)
initiate_document_version_upload(parent_folder_id, params::Dict{String,<:Any})

Creates a new document object and version object. The client specifies the parent folder ID and name of the document to upload. The ID is optionally specified when creating a new version of an existing document. This is the first step to upload a document. Next, upload the document to the URL returned from the call, and then call UpdateDocumentVersion. To cancel the document upload, call AbortDocumentVersionUpload.

Arguments

parent_folder_id: The ID of the parent folder.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"ContentCreatedTimestamp": The timestamp when the content of the document was originally created.

"ContentModifiedTimestamp": The timestamp when the content of the document was modified.

"ContentType": The content type of the document.

"DocumentSizeInBytes": The size of the document, in bytes.

"Id": The ID of the document.

"Name": The name of the document.

Main.Workdocs.remove_all_resource_permissions — Method

remove_all_resource_permissions(resource_id)
remove_all_resource_permissions(resource_id, params::Dict{String,<:Any})

Removes all the permissions from the specified resource.

Arguments

resource_id: The ID of the resource.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

Main.Workdocs.remove_resource_permission — Method

remove_resource_permission(principal_id, resource_id)
remove_resource_permission(principal_id, resource_id, params::Dict{String,<:Any})

Removes the permission for the specified principal from the specified resource.

Arguments

principal_id: The principal ID of the resource.

resource_id: The ID of the resource.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"type": The principal type of the resource.

Main.Workdocs.update_document — Method

update_document(document_id)
update_document(document_id, params::Dict{String,<:Any})

Updates the specified attributes of a document. The user must have access to both the document and its parent folder, if applicable.

Arguments

document_id: The ID of the document.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"Name": The name of the document.

"ParentFolderId": The ID of the parent folder.

"ResourceState": The resource state of the document. Only ACTIVE and RECYCLED are supported.

Main.Workdocs.update_document_version — Method

update_document_version(document_id, version_id)
update_document_version(document_id, version_id, params::Dict{String,<:Any})

Changes the status of the document version to ACTIVE. Amazon WorkDocs also sets its document container to ACTIVE. This is the last step in a document upload, after the client uploads the document to an S3-presigned URL returned by InitiateDocumentVersionUpload.

Arguments

document_id: The ID of the document.

version_id: The version ID of the document.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"VersionStatus": The status of the version.

Main.Workdocs.update_folder — Method

update_folder(folder_id)
update_folder(folder_id, params::Dict{String,<:Any})

Updates the specified attributes of the specified folder. The user must have access to both the folder and its parent folder, if applicable.

Arguments

folder_id: The ID of the folder.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"Name": The name of the folder.

"ParentFolderId": The ID of the parent folder.

"ResourceState": The resource state of the folder. Only ACTIVE and RECYCLED are accepted values from the API.

Main.Workdocs.update_user — Method

update_user(user_id)
update_user(user_id, params::Dict{String,<:Any})

Updates the specified attributes of the specified user, and grants or revokes administrative privileges to the Amazon WorkDocs site.

Arguments

user_id: The ID of the user.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Authentication": Amazon WorkDocs authentication token. Not required when using AWS administrator credentials to access the API.

"GivenName": The given name of the user.

"GrantPoweruserPrivileges": Boolean value to determine whether the user is granted Poweruser privileges.

"Locale": The locale of the user.

"StorageRule": The amount of storage for the user.

"Surname": The surname of the user.

"TimeZoneId": The time zone ID of the user.

"Type": The type of the user.