Worklink

This page documents function available when using the Worklink module, created with @service Worklink.

Index

Main.Worklink.associate_domain

Main.Worklink.associate_website_authorization_provider

Main.Worklink.associate_website_certificate_authority

Main.Worklink.create_fleet

Main.Worklink.delete_fleet

Main.Worklink.describe_audit_stream_configuration

Main.Worklink.describe_company_network_configuration

Main.Worklink.describe_device

Main.Worklink.describe_device_policy_configuration

Main.Worklink.describe_domain

Main.Worklink.describe_fleet_metadata

Main.Worklink.describe_identity_provider_configuration

Main.Worklink.describe_website_certificate_authority

Main.Worklink.disassociate_domain

Main.Worklink.disassociate_website_authorization_provider

Main.Worklink.disassociate_website_certificate_authority

Main.Worklink.list_devices

Main.Worklink.list_domains

Main.Worklink.list_fleets

Main.Worklink.list_tags_for_resource

Main.Worklink.list_website_authorization_providers

Main.Worklink.list_website_certificate_authorities

Main.Worklink.restore_domain_access

Main.Worklink.revoke_domain_access

Main.Worklink.sign_out_user

Main.Worklink.tag_resource

Main.Worklink.untag_resource

Main.Worklink.update_audit_stream_configuration

Main.Worklink.update_company_network_configuration

Main.Worklink.update_device_policy_configuration

Main.Worklink.update_domain_metadata

Main.Worklink.update_fleet_metadata

Main.Worklink.update_identity_provider_configuration

Documentation

Main.Worklink.associate_domain — Method

associate_domain(acm_certificate_arn, domain_name, fleet_arn)
associate_domain(acm_certificate_arn, domain_name, fleet_arn, params::Dict{String,<:Any})

Specifies a domain to be associated to Amazon WorkLink.

Arguments

acm_certificate_arn: The ARN of an issued ACM certificate that is valid for the domain being associated.

domain_name: The fully qualified domain name (FQDN).

fleet_arn: The Amazon Resource Name (ARN) of the fleet.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DisplayName": The name to display.

Main.Worklink.associate_website_authorization_provider — Method

associate_website_authorization_provider(authorization_provider_type, fleet_arn)
associate_website_authorization_provider(authorization_provider_type, fleet_arn, params::Dict{String,<:Any})

Associates a website authorization provider with a specified fleet. This is used to authorize users against associated websites in the company network.

Arguments

authorization_provider_type: The authorization provider type.

fleet_arn: The ARN of the fleet.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DomainName": The domain name of the authorization provider. This applies only to SAML-based authorization providers.

Main.Worklink.associate_website_certificate_authority — Method

associate_website_certificate_authority(certificate, fleet_arn)
associate_website_certificate_authority(certificate, fleet_arn, params::Dict{String,<:Any})

Imports the root certificate of a certificate authority (CA) used to obtain TLS certificates used by associated websites within the company network.

Arguments

certificate: The root certificate of the CA.

fleet_arn: The ARN of the fleet.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DisplayName": The certificate name to display.

Main.Worklink.create_fleet — Method

create_fleet(fleet_name)
create_fleet(fleet_name, params::Dict{String,<:Any})

Creates a fleet. A fleet consists of resources and the configuration that delivers associated websites to authorized users who download and set up the Amazon WorkLink app.

Arguments

fleet_name: A unique name for the fleet.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DisplayName": The fleet name to display.

"OptimizeForEndUserLocation": The option to optimize for better performance by routing traffic through the closest AWS Region to users, which may be outside of your home Region.

"Tags": The tags to add to the resource. A tag is a key-value pair.

Main.Worklink.delete_fleet — Method

delete_fleet(fleet_arn)
delete_fleet(fleet_arn, params::Dict{String,<:Any})

Deletes a fleet. Prevents users from accessing previously associated websites.

Arguments

fleet_arn: The ARN of the fleet.

Main.Worklink.describe_audit_stream_configuration — Method

describe_audit_stream_configuration(fleet_arn)
describe_audit_stream_configuration(fleet_arn, params::Dict{String,<:Any})

Describes the configuration for delivering audit streams to the customer account.

Arguments

fleet_arn: The ARN of the fleet.

Main.Worklink.describe_company_network_configuration — Method

describe_company_network_configuration(fleet_arn)
describe_company_network_configuration(fleet_arn, params::Dict{String,<:Any})

Describes the networking configuration to access the internal websites associated with the specified fleet.

Arguments

fleet_arn: The ARN of the fleet.

Main.Worklink.describe_device — Method

describe_device(device_id, fleet_arn)
describe_device(device_id, fleet_arn, params::Dict{String,<:Any})

Provides information about a user's device.

Arguments

device_id: A unique identifier for a registered user's device.

fleet_arn: The ARN of the fleet.

Main.Worklink.describe_device_policy_configuration — Method

describe_device_policy_configuration(fleet_arn)
describe_device_policy_configuration(fleet_arn, params::Dict{String,<:Any})

Describes the device policy configuration for the specified fleet.

Arguments

fleet_arn: The ARN of the fleet.

Main.Worklink.describe_domain — Method

describe_domain(domain_name, fleet_arn)
describe_domain(domain_name, fleet_arn, params::Dict{String,<:Any})

Provides information about the domain.

Arguments

domain_name: The name of the domain.

fleet_arn: The ARN of the fleet.

Main.Worklink.describe_fleet_metadata — Method

describe_fleet_metadata(fleet_arn)
describe_fleet_metadata(fleet_arn, params::Dict{String,<:Any})

Provides basic information for the specified fleet, excluding identity provider, networking, and device configuration details.

Arguments

fleet_arn: The Amazon Resource Name (ARN) of the fleet.

Main.Worklink.describe_identity_provider_configuration — Method

describe_identity_provider_configuration(fleet_arn)
describe_identity_provider_configuration(fleet_arn, params::Dict{String,<:Any})

Describes the identity provider configuration of the specified fleet.

Arguments

fleet_arn: The ARN of the fleet.

Main.Worklink.describe_website_certificate_authority — Method

describe_website_certificate_authority(fleet_arn, website_ca_id)
describe_website_certificate_authority(fleet_arn, website_ca_id, params::Dict{String,<:Any})

Provides information about the certificate authority.

Arguments

fleet_arn: The ARN of the fleet.

website_ca_id: A unique identifier for the certificate authority.

Main.Worklink.disassociate_domain — Method

disassociate_domain(domain_name, fleet_arn)
disassociate_domain(domain_name, fleet_arn, params::Dict{String,<:Any})

Disassociates a domain from Amazon WorkLink. End users lose the ability to access the domain with Amazon WorkLink.

Arguments

domain_name: The name of the domain.

fleet_arn: The ARN of the fleet.

Main.Worklink.disassociate_website_authorization_provider — Method

disassociate_website_authorization_provider(authorization_provider_id, fleet_arn)
disassociate_website_authorization_provider(authorization_provider_id, fleet_arn, params::Dict{String,<:Any})

Disassociates a website authorization provider from a specified fleet. After the disassociation, users can't load any associated websites that require this authorization provider.

Arguments

authorization_provider_id: A unique identifier for the authorization provider.

fleet_arn: The ARN of the fleet.

Main.Worklink.disassociate_website_certificate_authority — Method

disassociate_website_certificate_authority(fleet_arn, website_ca_id)
disassociate_website_certificate_authority(fleet_arn, website_ca_id, params::Dict{String,<:Any})

Removes a certificate authority (CA).

Arguments

fleet_arn: The ARN of the fleet.

website_ca_id: A unique identifier for the CA.

Main.Worklink.list_devices — Method

list_devices(fleet_arn)
list_devices(fleet_arn, params::Dict{String,<:Any})

Retrieves a list of devices registered with the specified fleet.

Arguments

fleet_arn: The ARN of the fleet.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to be included in the next page.

"NextToken": The pagination token used to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

Main.Worklink.list_domains — Method

list_domains(fleet_arn)
list_domains(fleet_arn, params::Dict{String,<:Any})

Retrieves a list of domains associated to a specified fleet.

Arguments

fleet_arn: The ARN of the fleet.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to be included in the next page.

"NextToken": The pagination token used to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

Main.Worklink.list_fleets — Method

list_fleets()
list_fleets(params::Dict{String,<:Any})

Retrieves a list of fleets for the current account and Region.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to be included in the next page.

"NextToken": The pagination token used to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

Main.Worklink.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Retrieves a list of tags for the specified resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the fleet.

Main.Worklink.list_website_authorization_providers — Method

list_website_authorization_providers(fleet_arn)
list_website_authorization_providers(fleet_arn, params::Dict{String,<:Any})

Retrieves a list of website authorization providers associated with a specified fleet.

Arguments

fleet_arn: The ARN of the fleet.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to be included in the next page.

"NextToken": The pagination token to use to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

Main.Worklink.list_website_certificate_authorities — Method

list_website_certificate_authorities(fleet_arn)
list_website_certificate_authorities(fleet_arn, params::Dict{String,<:Any})

Retrieves a list of certificate authorities added for the current account and Region.

Arguments

fleet_arn: The ARN of the fleet.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to be included in the next page.

"NextToken": The pagination token used to retrieve the next page of results for this operation. If this value is null, it retrieves the first page.

Main.Worklink.restore_domain_access — Method

restore_domain_access(domain_name, fleet_arn)
restore_domain_access(domain_name, fleet_arn, params::Dict{String,<:Any})

Moves a domain to ACTIVE status if it was in the INACTIVE status.

Arguments

domain_name: The name of the domain.

fleet_arn: The ARN of the fleet.

Main.Worklink.revoke_domain_access — Method

revoke_domain_access(domain_name, fleet_arn)
revoke_domain_access(domain_name, fleet_arn, params::Dict{String,<:Any})

Moves a domain to INACTIVE status if it was in the ACTIVE status.

Arguments

domain_name: The name of the domain.

fleet_arn: The ARN of the fleet.

Main.Worklink.sign_out_user — Method

sign_out_user(fleet_arn, username)
sign_out_user(fleet_arn, username, params::Dict{String,<:Any})

Signs the user out from all of their devices. The user can sign in again if they have valid credentials.

Arguments

fleet_arn: The ARN of the fleet.

username: The name of the user.

Main.Worklink.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Adds or overwrites one or more tags for the specified resource, such as a fleet. Each tag consists of a key and an optional value. If a resource already has a tag with the same key, this operation updates its value.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the fleet.

tags: The tags to add to the resource. A tag is a key-value pair.

Main.Worklink.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes one or more tags from the specified resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the fleet.

tag_keys: The list of tag keys to remove from the resource.

Main.Worklink.update_audit_stream_configuration — Method

update_audit_stream_configuration(fleet_arn)
update_audit_stream_configuration(fleet_arn, params::Dict{String,<:Any})

Updates the audit stream configuration for the fleet.

Arguments

fleet_arn: The ARN of the fleet.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AuditStreamArn": The ARN of the Amazon Kinesis data stream that receives the audit events.

Main.Worklink.update_company_network_configuration — Method

update_company_network_configuration(fleet_arn, security_group_ids, subnet_ids, vpc_id)
update_company_network_configuration(fleet_arn, security_group_ids, subnet_ids, vpc_id, params::Dict{String,<:Any})

Updates the company network configuration for the fleet.

Arguments

fleet_arn: The ARN of the fleet.

security_group_ids: The security groups associated with access to the provided subnets.

subnet_ids: The subnets used for X-ENI connections from Amazon WorkLink rendering containers.

vpc_id: The VPC with connectivity to associated websites.

Main.Worklink.update_device_policy_configuration — Method

update_device_policy_configuration(fleet_arn)
update_device_policy_configuration(fleet_arn, params::Dict{String,<:Any})

Updates the device policy configuration for the fleet.

Arguments

fleet_arn: The ARN of the fleet.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DeviceCaCertificate": The certificate chain, including intermediate certificates and the root certificate authority certificate used to issue device certificates.

Main.Worklink.update_domain_metadata — Method

update_domain_metadata(domain_name, fleet_arn)
update_domain_metadata(domain_name, fleet_arn, params::Dict{String,<:Any})

Updates domain metadata, such as DisplayName.

Arguments

domain_name: The name of the domain.

fleet_arn: The ARN of the fleet.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DisplayName": The name to display.

Main.Worklink.update_fleet_metadata — Method

update_fleet_metadata(fleet_arn)
update_fleet_metadata(fleet_arn, params::Dict{String,<:Any})

Updates fleet metadata, such as DisplayName.

Arguments

fleet_arn: The ARN of the fleet.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DisplayName": The fleet name to display. The existing DisplayName is unset if null is passed.

"OptimizeForEndUserLocation": The option to optimize for better performance by routing traffic through the closest AWS Region to users, which may be outside of your home Region.

Main.Worklink.update_identity_provider_configuration — Method

update_identity_provider_configuration(fleet_arn, identity_provider_type)
update_identity_provider_configuration(fleet_arn, identity_provider_type, params::Dict{String,<:Any})

Updates the identity provider configuration for the fleet.

Arguments

fleet_arn: The ARN of the fleet.

identity_provider_type: The type of identity provider.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"IdentityProviderSamlMetadata": The SAML metadata document provided by the customer’s identity provider. The existing IdentityProviderSamlMetadata is unset if null is passed.