Workspaces Thin Client

This page documents function available when using the Workspaces_Thin_Client module, created with @service Workspaces_Thin_Client.

Index

Main.Workspaces_Thin_Client.create_environment

Main.Workspaces_Thin_Client.delete_device

Main.Workspaces_Thin_Client.delete_environment

Main.Workspaces_Thin_Client.deregister_device

Main.Workspaces_Thin_Client.get_device

Main.Workspaces_Thin_Client.get_environment

Main.Workspaces_Thin_Client.get_software_set

Main.Workspaces_Thin_Client.list_devices

Main.Workspaces_Thin_Client.list_environments

Main.Workspaces_Thin_Client.list_software_sets

Main.Workspaces_Thin_Client.list_tags_for_resource

Main.Workspaces_Thin_Client.tag_resource

Main.Workspaces_Thin_Client.untag_resource

Main.Workspaces_Thin_Client.update_device

Main.Workspaces_Thin_Client.update_environment

Main.Workspaces_Thin_Client.update_software_set

Documentation

Main.Workspaces_Thin_Client.create_environment — Method

create_environment(desktop_arn)
create_environment(desktop_arn, params::Dict{String,<:Any})

Creates an environment for your thin client devices.

Arguments

desktop_arn: The Amazon Resource Name (ARN) of the desktop to stream from Amazon WorkSpaces, WorkSpaces Web, or AppStream 2.0.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Specifies a unique, case-sensitive identifier that you provide to ensure the idempotency of the request. This lets you safely retry the request without accidentally performing the same operation a second time. Passing the same value to a later call to an operation requires that you also pass the same value for all other parameters. We recommend that you use a UUID type of value. If you don't provide this value, then Amazon Web Services generates a random one for you. If you retry the operation with the same ClientToken, but with different parameters, the retry fails with an IdempotentParameterMismatch error.

"desiredSoftwareSetId": The ID of the software set to apply.

"desktopEndpoint": The URL for the identity provider login (only for environments that use AppStream 2.0).

"deviceCreationTags": A map of the key-value pairs of the tag or tags to assign to the newly created devices for this environment.

"kmsKeyArn": The Amazon Resource Name (ARN) of the Key Management Service key to use to encrypt the environment.

"maintenanceWindow": A specification for a time window to apply software updates.

"name": The name for the environment.

"softwareSetUpdateMode": An option to define which software updates to apply.

"softwareSetUpdateSchedule": An option to define if software updates should be applied within a maintenance window.

"tags": A map of the key-value pairs of the tag or tags to assign to the resource.

Main.Workspaces_Thin_Client.delete_device — Method

delete_device(id)
delete_device(id, params::Dict{String,<:Any})

Deletes a thin client device.

Arguments

id: The ID of the device to delete.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Specifies a unique, case-sensitive identifier that you provide to ensure the idempotency of the request. This lets you safely retry the request without accidentally performing the same operation a second time. Passing the same value to a later call to an operation requires that you also pass the same value for all other parameters. We recommend that you use a UUID type of value. If you don't provide this value, then Amazon Web Services generates a random one for you. If you retry the operation with the same ClientToken, but with different parameters, the retry fails with an IdempotentParameterMismatch error.

Main.Workspaces_Thin_Client.delete_environment — Method

delete_environment(id)
delete_environment(id, params::Dict{String,<:Any})

Deletes an environment.

Arguments

id: The ID of the environment to delete.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Specifies a unique, case-sensitive identifier that you provide to ensure the idempotency of the request. This lets you safely retry the request without accidentally performing the same operation a second time. Passing the same value to a later call to an operation requires that you also pass the same value for all other parameters. We recommend that you use a UUID type of value. If you don't provide this value, then Amazon Web Services generates a random one for you. If you retry the operation with the same ClientToken, but with different parameters, the retry fails with an IdempotentParameterMismatch error.

Main.Workspaces_Thin_Client.deregister_device — Method

deregister_device(id)
deregister_device(id, params::Dict{String,<:Any})

Deregisters a thin client device.

Arguments

id: The ID of the device to deregister.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Specifies a unique, case-sensitive identifier that you provide to ensure the idempotency of the request. This lets you safely retry the request without accidentally performing the same operation a second time. Passing the same value to a later call to an operation requires that you also pass the same value for all other parameters. We recommend that you use a UUID type of value. If you don't provide this value, then Amazon Web Services generates a random one for you. If you retry the operation with the same ClientToken, but with different parameters, the retry fails with an IdempotentParameterMismatch error.

"targetDeviceStatus": The desired new status for the device.

Main.Workspaces_Thin_Client.get_device — Method

get_device(id)
get_device(id, params::Dict{String,<:Any})

Returns information for a thin client device.

Arguments

id: The ID of the device for which to return information.

Main.Workspaces_Thin_Client.get_environment — Method

get_environment(id)
get_environment(id, params::Dict{String,<:Any})

Returns information for an environment.

Arguments

id: The ID of the environment for which to return information.

Main.Workspaces_Thin_Client.get_software_set — Method

get_software_set(id)
get_software_set(id, params::Dict{String,<:Any})

Returns information for a software set.

Arguments

id: The ID of the software set for which to return information.

Main.Workspaces_Thin_Client.list_devices — Method

list_devices()
list_devices(params::Dict{String,<:Any})

Returns a list of thin client devices.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results that are returned per call. You can use nextToken to obtain further pages of results. This is only an upper limit. The actual number of results returned per call might be fewer than the specified maximum.

"nextToken": If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

Main.Workspaces_Thin_Client.list_environments — Method

list_environments()
list_environments(params::Dict{String,<:Any})

Returns a list of environments.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results that are returned per call. You can use nextToken to obtain further pages of results. This is only an upper limit. The actual number of results returned per call might be fewer than the specified maximum.

"nextToken": If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

Main.Workspaces_Thin_Client.list_software_sets — Method

list_software_sets()
list_software_sets(params::Dict{String,<:Any})

Returns a list of software sets.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results that are returned per call. You can use nextToken to obtain further pages of results. This is only an upper limit. The actual number of results returned per call might be fewer than the specified maximum.

"nextToken": If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

Main.Workspaces_Thin_Client.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Returns a list of tags for a resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource for which you want to retrieve tags.

Main.Workspaces_Thin_Client.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Assigns one or more tags (key-value pairs) to the specified resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource that you want to tag.

tags: A map of the key-value pairs of the tag or tags to assign to the resource.

Main.Workspaces_Thin_Client.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes a tag or tags from a resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource that you want to untag.

tag_keys: The keys of the key-value pairs for the tag or tags you want to remove from the specified resource.

Main.Workspaces_Thin_Client.update_device — Method

update_device(id)
update_device(id, params::Dict{String,<:Any})

Updates a thin client device.

Arguments

id: The ID of the device to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"desiredSoftwareSetId": The ID of the software set to apply.

"name": The name of the device to update.

"softwareSetUpdateSchedule": An option to define if software updates should be applied within a maintenance window.

Main.Workspaces_Thin_Client.update_environment — Method

update_environment(id)
update_environment(id, params::Dict{String,<:Any})

Updates an environment.

Arguments

id: The ID of the environment to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"desiredSoftwareSetId": The ID of the software set to apply.

"desktopArn": The Amazon Resource Name (ARN) of the desktop to stream from Amazon WorkSpaces, WorkSpaces Web, or AppStream 2.0.

"desktopEndpoint": The URL for the identity provider login (only for environments that use AppStream 2.0).

"deviceCreationTags": A map of the key-value pairs of the tag or tags to assign to the newly created devices for this environment.

"maintenanceWindow": A specification for a time window to apply software updates.

"name": The name of the environment to update.

"softwareSetUpdateMode": An option to define which software updates to apply.

"softwareSetUpdateSchedule": An option to define if software updates should be applied within a maintenance window.

Main.Workspaces_Thin_Client.update_software_set — Method

update_software_set(id, validation_status)
update_software_set(id, validation_status, params::Dict{String,<:Any})

Updates a software set.

Arguments

id: The ID of the software set to update.

validation_status: An option to define if the software set has been validated.