Workspaces Web

This page documents function available when using the Workspaces_Web module, created with @service Workspaces_Web.

Index

Main.Workspaces_Web.associate_browser_settings

Main.Workspaces_Web.associate_network_settings

Main.Workspaces_Web.associate_trust_store

Main.Workspaces_Web.associate_user_access_logging_settings

Main.Workspaces_Web.associate_user_settings

Main.Workspaces_Web.create_browser_settings

Main.Workspaces_Web.create_identity_provider

Main.Workspaces_Web.create_network_settings

Main.Workspaces_Web.create_portal

Main.Workspaces_Web.create_trust_store

Main.Workspaces_Web.create_user_access_logging_settings

Main.Workspaces_Web.create_user_settings

Main.Workspaces_Web.delete_browser_settings

Main.Workspaces_Web.delete_identity_provider

Main.Workspaces_Web.delete_network_settings

Main.Workspaces_Web.delete_portal

Main.Workspaces_Web.delete_trust_store

Main.Workspaces_Web.delete_user_access_logging_settings

Main.Workspaces_Web.delete_user_settings

Main.Workspaces_Web.disassociate_browser_settings

Main.Workspaces_Web.disassociate_network_settings

Main.Workspaces_Web.disassociate_trust_store

Main.Workspaces_Web.disassociate_user_access_logging_settings

Main.Workspaces_Web.disassociate_user_settings

Main.Workspaces_Web.get_browser_settings

Main.Workspaces_Web.get_identity_provider

Main.Workspaces_Web.get_network_settings

Main.Workspaces_Web.get_portal

Main.Workspaces_Web.get_portal_service_provider_metadata

Main.Workspaces_Web.get_trust_store

Main.Workspaces_Web.get_trust_store_certificate

Main.Workspaces_Web.get_user_access_logging_settings

Main.Workspaces_Web.get_user_settings

Main.Workspaces_Web.list_browser_settings

Main.Workspaces_Web.list_identity_providers

Main.Workspaces_Web.list_network_settings

Main.Workspaces_Web.list_portals

Main.Workspaces_Web.list_tags_for_resource

Main.Workspaces_Web.list_trust_store_certificates

Main.Workspaces_Web.list_trust_stores

Main.Workspaces_Web.list_user_access_logging_settings

Main.Workspaces_Web.list_user_settings

Main.Workspaces_Web.tag_resource

Main.Workspaces_Web.untag_resource

Main.Workspaces_Web.update_browser_settings

Main.Workspaces_Web.update_identity_provider

Main.Workspaces_Web.update_network_settings

Main.Workspaces_Web.update_portal

Main.Workspaces_Web.update_trust_store

Main.Workspaces_Web.update_user_access_logging_settings

Main.Workspaces_Web.update_user_settings

Documentation

Main.Workspaces_Web.associate_browser_settings — Method

associate_browser_settings(browser_settings_arn, portal_arn)
associate_browser_settings(browser_settings_arn, portal_arn, params::Dict{String,<:Any})

Associates a browser settings resource with a web portal.

Arguments

browser_settings_arn: The ARN of the browser settings.

portal_arn: The ARN of the web portal.

Main.Workspaces_Web.associate_network_settings — Method

associate_network_settings(network_settings_arn, portal_arn)
associate_network_settings(network_settings_arn, portal_arn, params::Dict{String,<:Any})

Associates a network settings resource with a web portal.

Arguments

network_settings_arn: The ARN of the network settings.

portal_arn: The ARN of the web portal.

Main.Workspaces_Web.associate_trust_store — Method

associate_trust_store(portal_arn, trust_store_arn)
associate_trust_store(portal_arn, trust_store_arn, params::Dict{String,<:Any})

Associates a trust store with a web portal.

Arguments

portal_arn: The ARN of the web portal.

trust_store_arn: The ARN of the trust store.

Main.Workspaces_Web.associate_user_access_logging_settings — Method

associate_user_access_logging_settings(portal_arn, user_access_logging_settings_arn)
associate_user_access_logging_settings(portal_arn, user_access_logging_settings_arn, params::Dict{String,<:Any})

Associates a user access logging settings resource with a web portal.

Arguments

portal_arn: The ARN of the web portal.

user_access_logging_settings_arn: The ARN of the user access logging settings.

Main.Workspaces_Web.associate_user_settings — Method

associate_user_settings(portal_arn, user_settings_arn)
associate_user_settings(portal_arn, user_settings_arn, params::Dict{String,<:Any})

Associates a user settings resource with a web portal.

Arguments

portal_arn: The ARN of the web portal.

user_settings_arn: The ARN of the user settings.

Main.Workspaces_Web.create_browser_settings — Method

create_browser_settings(browser_policy)
create_browser_settings(browser_policy, params::Dict{String,<:Any})

Creates a browser settings resource that can be associated with a web portal. Once associated with a web portal, browser settings control how the browser will behave once a user starts a streaming session for the web portal.

Arguments

browser_policy: A JSON string containing Chrome Enterprise policies that will be applied to all streaming sessions.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"additionalEncryptionContext": Additional encryption context of the browser settings.

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, subsequent retries with the same client token returns the result from the original successful request. If you do not specify a client token, one is automatically generated by the AWS SDK.

"customerManagedKey": The custom managed key of the browser settings.

"tags": The tags to add to the browser settings resource. A tag is a key-value pair.

Main.Workspaces_Web.create_identity_provider — Method

create_identity_provider(identity_provider_details, identity_provider_name, identity_provider_type, portal_arn)
create_identity_provider(identity_provider_details, identity_provider_name, identity_provider_type, portal_arn, params::Dict{String,<:Any})

Creates an identity provider resource that is then associated with a web portal.

Arguments

identity_provider_details: The identity provider details. The following list describes the provider detail keys for each identity provider type. For Google and Login with Amazon: clientid clientsecret authorizescopes For Facebook: clientid clientsecret authorizescopes apiversion For Sign in with Apple: clientid teamid keyid privatekey authorizescopes For OIDC providers: clientid clientsecret attributesrequestmethod oidcissuer authorizescopes authorizeurl if not available from discovery URL specified by oidcissuer key tokenurl if not available from discovery URL specified by oidcissuer key attributesurl if not available from discovery URL specified by oidcissuer key jwksuri if not available from discovery URL specified by oidcissuer key For SAML providers: MetadataFile OR MetadataURL IDPSignout (boolean) optional

identity_provider_name: The identity provider name.

identity_provider_type: The identity provider type.

portal_arn: The ARN of the web portal.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, subsequent retries with the same client token returns the result from the original successful request. If you do not specify a client token, one is automatically generated by the AWS SDK.

Main.Workspaces_Web.create_network_settings — Method

create_network_settings(security_group_ids, subnet_ids, vpc_id)
create_network_settings(security_group_ids, subnet_ids, vpc_id, params::Dict{String,<:Any})

Creates a network settings resource that can be associated with a web portal. Once associated with a web portal, network settings define how streaming instances will connect with your specified VPC.

Arguments

security_group_ids: One or more security groups used to control access from streaming instances to your VPC.

subnet_ids: The subnets in which network interfaces are created to connect streaming instances to your VPC. At least two of these subnets must be in different availability zones.

vpc_id: The VPC that streaming instances will connect to.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, subsequent retries with the same client token returns the result from the original successful request. If you do not specify a client token, one is automatically generated by the AWS SDK.

"tags": The tags to add to the network settings resource. A tag is a key-value pair.

Main.Workspaces_Web.create_portal — Method

create_portal()
create_portal(params::Dict{String,<:Any})

Creates a web portal.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"additionalEncryptionContext": The additional encryption context of the portal.

"authenticationType": The type of authentication integration points used when signing into the web portal. Defaults to Standard. Standard web portals are authenticated directly through your identity provider. You need to call CreateIdentityProvider to integrate your identity provider with your web portal. User and group access to your web portal is controlled through your identity provider. IAMIdentityCenter web portals are authenticated through AWS IAM Identity Center (successor to AWS Single Sign-On). They provide additional features, such as IdP-initiated authentication. Identity sources (including external identity provider integration), plus user and group access to your web portal, can be configured in the IAM Identity Center.

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, subsequent retries with the same client token returns the result from the original successful request. If you do not specify a client token, one is automatically generated by the AWS SDK.

"customerManagedKey": The customer managed key of the web portal.

"displayName": The name of the web portal. This is not visible to users who log into the web portal.

"tags": The tags to add to the web portal. A tag is a key-value pair.

Main.Workspaces_Web.create_trust_store — Method

create_trust_store(certificate_list)
create_trust_store(certificate_list, params::Dict{String,<:Any})

Creates a trust store that can be associated with a web portal. A trust store contains certificate authority (CA) certificates. Once associated with a web portal, the browser in a streaming session will recognize certificates that have been issued using any of the CAs in the trust store. If your organization has internal websites that use certificates issued by private CAs, you should add the private CA certificate to the trust store.

Arguments

certificate_list: A list of CA certificates to be added to the trust store.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, subsequent retries with the same client token returns the result from the original successful request. If you do not specify a client token, one is automatically generated by the AWS SDK.

"tags": The tags to add to the trust store. A tag is a key-value pair.

Main.Workspaces_Web.create_user_access_logging_settings — Method

create_user_access_logging_settings(kinesis_stream_arn)
create_user_access_logging_settings(kinesis_stream_arn, params::Dict{String,<:Any})

Creates a user access logging settings resource that can be associated with a web portal.

Arguments

kinesis_stream_arn: The ARN of the Kinesis stream.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, subsequent retries with the same client token returns the result from the original successful request. If you do not specify a client token, one is automatically generated by the AWS SDK.

"tags": The tags to add to the user settings resource. A tag is a key-value pair.

Main.Workspaces_Web.create_user_settings — Method

create_user_settings(copy_allowed, download_allowed, paste_allowed, print_allowed, upload_allowed)
create_user_settings(copy_allowed, download_allowed, paste_allowed, print_allowed, upload_allowed, params::Dict{String,<:Any})

Creates a user settings resource that can be associated with a web portal. Once associated with a web portal, user settings control how users can transfer data between a streaming session and the their local devices.

Arguments

copy_allowed: Specifies whether the user can copy text from the streaming session to the local device.

download_allowed: Specifies whether the user can download files from the streaming session to the local device.

paste_allowed: Specifies whether the user can paste text from the local device to the streaming session.

print_allowed: Specifies whether the user can print to the local device.

upload_allowed: Specifies whether the user can upload files from the local device to the streaming session.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, subsequent retries with the same client token returns the result from the original successful request. If you do not specify a client token, one is automatically generated by the AWS SDK.

"disconnectTimeoutInMinutes": The amount of time that a streaming session remains active after users disconnect.

"idleDisconnectTimeoutInMinutes": The amount of time that users can be idle (inactive) before they are disconnected from their streaming session and the disconnect timeout interval begins.

"tags": The tags to add to the user settings resource. A tag is a key-value pair.

Main.Workspaces_Web.delete_browser_settings — Method

delete_browser_settings(browser_settings_arn)
delete_browser_settings(browser_settings_arn, params::Dict{String,<:Any})

Deletes browser settings.

Arguments

browser_settings_arn: The ARN of the browser settings.

Main.Workspaces_Web.delete_identity_provider — Method

delete_identity_provider(identity_provider_arn)
delete_identity_provider(identity_provider_arn, params::Dict{String,<:Any})

Deletes the identity provider.

Arguments

identity_provider_arn: The ARN of the identity provider.

Main.Workspaces_Web.delete_network_settings — Method

delete_network_settings(network_settings_arn)
delete_network_settings(network_settings_arn, params::Dict{String,<:Any})

Deletes network settings.

Arguments

network_settings_arn: The ARN of the network settings.

Main.Workspaces_Web.delete_portal — Method

delete_portal(portal_arn)
delete_portal(portal_arn, params::Dict{String,<:Any})

Deletes a web portal.

Arguments

portal_arn: The ARN of the web portal.

Main.Workspaces_Web.delete_trust_store — Method

delete_trust_store(trust_store_arn)
delete_trust_store(trust_store_arn, params::Dict{String,<:Any})

Deletes the trust store.

Arguments

trust_store_arn: The ARN of the trust store.

Main.Workspaces_Web.delete_user_access_logging_settings — Method

delete_user_access_logging_settings(user_access_logging_settings_arn)
delete_user_access_logging_settings(user_access_logging_settings_arn, params::Dict{String,<:Any})

Deletes user access logging settings.

Arguments

user_access_logging_settings_arn: The ARN of the user access logging settings.

Main.Workspaces_Web.delete_user_settings — Method

delete_user_settings(user_settings_arn)
delete_user_settings(user_settings_arn, params::Dict{String,<:Any})

Deletes user settings.

Arguments

user_settings_arn: The ARN of the user settings.

Main.Workspaces_Web.disassociate_browser_settings — Method

disassociate_browser_settings(portal_arn)
disassociate_browser_settings(portal_arn, params::Dict{String,<:Any})

Disassociates browser settings from a web portal.

Arguments

portal_arn: The ARN of the web portal.

Main.Workspaces_Web.disassociate_network_settings — Method

disassociate_network_settings(portal_arn)
disassociate_network_settings(portal_arn, params::Dict{String,<:Any})

Disassociates network settings from a web portal.

Arguments

portal_arn: The ARN of the web portal.

Main.Workspaces_Web.disassociate_trust_store — Method

disassociate_trust_store(portal_arn)
disassociate_trust_store(portal_arn, params::Dict{String,<:Any})

Disassociates a trust store from a web portal.

Arguments

portal_arn: The ARN of the web portal.

Main.Workspaces_Web.disassociate_user_access_logging_settings — Method

disassociate_user_access_logging_settings(portal_arn)
disassociate_user_access_logging_settings(portal_arn, params::Dict{String,<:Any})

Disassociates user access logging settings from a web portal.

Arguments

portal_arn: The ARN of the web portal.

Main.Workspaces_Web.disassociate_user_settings — Method

disassociate_user_settings(portal_arn)
disassociate_user_settings(portal_arn, params::Dict{String,<:Any})

Disassociates user settings from a web portal.

Arguments

portal_arn: The ARN of the web portal.

Main.Workspaces_Web.get_browser_settings — Method

get_browser_settings(browser_settings_arn)
get_browser_settings(browser_settings_arn, params::Dict{String,<:Any})

Gets browser settings.

Arguments

browser_settings_arn: The ARN of the browser settings.

Main.Workspaces_Web.get_identity_provider — Method

get_identity_provider(identity_provider_arn)
get_identity_provider(identity_provider_arn, params::Dict{String,<:Any})

Gets the identity provider.

Arguments

identity_provider_arn: The ARN of the identity provider.

Main.Workspaces_Web.get_network_settings — Method

get_network_settings(network_settings_arn)
get_network_settings(network_settings_arn, params::Dict{String,<:Any})

Gets the network settings.

Arguments

network_settings_arn: The ARN of the network settings.

Main.Workspaces_Web.get_portal — Method

get_portal(portal_arn)
get_portal(portal_arn, params::Dict{String,<:Any})

Gets the web portal.

Arguments

portal_arn: The ARN of the web portal.

Main.Workspaces_Web.get_portal_service_provider_metadata — Method

get_portal_service_provider_metadata(portal_arn)
get_portal_service_provider_metadata(portal_arn, params::Dict{String,<:Any})

Gets the service provider metadata.

Arguments

portal_arn: The ARN of the web portal.

Main.Workspaces_Web.get_trust_store — Method

get_trust_store(trust_store_arn)
get_trust_store(trust_store_arn, params::Dict{String,<:Any})

Gets the trust store.

Arguments

trust_store_arn: The ARN of the trust store.

Main.Workspaces_Web.get_trust_store_certificate — Method

get_trust_store_certificate(thumbprint, trust_store_arn)
get_trust_store_certificate(thumbprint, trust_store_arn, params::Dict{String,<:Any})

Gets the trust store certificate.

Arguments

thumbprint: The thumbprint of the trust store certificate.

trust_store_arn: The ARN of the trust store certificate.

Main.Workspaces_Web.get_user_access_logging_settings — Method

get_user_access_logging_settings(user_access_logging_settings_arn)
get_user_access_logging_settings(user_access_logging_settings_arn, params::Dict{String,<:Any})

Gets user access logging settings.

Arguments

user_access_logging_settings_arn: The ARN of the user access logging settings.

Main.Workspaces_Web.get_user_settings — Method

get_user_settings(user_settings_arn)
get_user_settings(user_settings_arn, params::Dict{String,<:Any})

Gets user settings.

Arguments

user_settings_arn: The ARN of the user settings.

Main.Workspaces_Web.list_browser_settings — Method

list_browser_settings()
list_browser_settings(params::Dict{String,<:Any})

Retrieves a list of browser settings.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to be included in the next page.

"nextToken": The pagination token used to retrieve the next page of results for this operation.

Main.Workspaces_Web.list_identity_providers — Method

list_identity_providers(portal_arn)
list_identity_providers(portal_arn, params::Dict{String,<:Any})

Retrieves a list of identity providers for a specific web portal.

Arguments

portal_arn: The ARN of the web portal.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to be included in the next page.

"nextToken": The pagination token used to retrieve the next page of results for this operation.

Main.Workspaces_Web.list_network_settings — Method

list_network_settings()
list_network_settings(params::Dict{String,<:Any})

Retrieves a list of network settings.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to be included in the next page.

"nextToken": The pagination token used to retrieve the next page of results for this operation.

Main.Workspaces_Web.list_portals — Method

list_portals()
list_portals(params::Dict{String,<:Any})

Retrieves a list or web portals.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to be included in the next page.

"nextToken": The pagination token used to retrieve the next page of results for this operation.

Main.Workspaces_Web.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Retrieves a list of tags for a resource.

Arguments

resource_arn: The ARN of the resource.

Main.Workspaces_Web.list_trust_store_certificates — Method

list_trust_store_certificates(trust_store_arn)
list_trust_store_certificates(trust_store_arn, params::Dict{String,<:Any})

Retrieves a list of trust store certificates.

Arguments

trust_store_arn: The ARN of the trust store

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to be included in the next page.

"nextToken": The pagination token used to retrieve the next page of results for this operation.

Main.Workspaces_Web.list_trust_stores — Method

list_trust_stores()
list_trust_stores(params::Dict{String,<:Any})

Retrieves a list of trust stores.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to be included in the next page.

"nextToken": The pagination token used to retrieve the next page of results for this operation.

Main.Workspaces_Web.list_user_access_logging_settings — Method

list_user_access_logging_settings()
list_user_access_logging_settings(params::Dict{String,<:Any})

Retrieves a list of user access logging settings.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to be included in the next page.

"nextToken": The pagination token used to retrieve the next page of results for this operation.

Main.Workspaces_Web.list_user_settings — Method

list_user_settings()
list_user_settings(params::Dict{String,<:Any})

Retrieves a list of user settings.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to be included in the next page.

"nextToken": The pagination token used to retrieve the next page of results for this operation.

Main.Workspaces_Web.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Adds or overwrites one or more tags for the specified resource.

Arguments

resource_arn: The ARN of the resource.

tags: The tags of the resource.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, subsequent retries with the same client token returns the result from the original successful request. If you do not specify a client token, one is automatically generated by the AWS SDK.

Main.Workspaces_Web.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes one or more tags from the specified resource.

Arguments

resource_arn: The ARN of the resource.

tag_keys: The list of tag keys to remove from the resource.

Main.Workspaces_Web.update_browser_settings — Method

update_browser_settings(browser_settings_arn)
update_browser_settings(browser_settings_arn, params::Dict{String,<:Any})

Updates browser settings.

Arguments

browser_settings_arn: The ARN of the browser settings.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"browserPolicy": A JSON string containing Chrome Enterprise policies that will be applied to all streaming sessions.

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, subsequent retries with the same client token return the result from the original successful request. If you do not specify a client token, one is automatically generated by the AWS SDK.

Main.Workspaces_Web.update_identity_provider — Method

update_identity_provider(identity_provider_arn)
update_identity_provider(identity_provider_arn, params::Dict{String,<:Any})

Updates the identity provider.

Arguments

identity_provider_arn: The ARN of the identity provider.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, subsequent retries with the same client token return the result from the original successful request. If you do not specify a client token, one is automatically generated by the AWS SDK.

"identityProviderDetails": The details of the identity provider. The following list describes the provider detail keys for each identity provider type. For Google and Login with Amazon: clientid clientsecret authorizescopes For Facebook: clientid clientsecret authorizescopes apiversion For Sign in with Apple: clientid teamid keyid privatekey authorizescopes For OIDC providers: clientid clientsecret attributesrequestmethod oidcissuer authorizescopes authorizeurl if not available from discovery URL specified by oidcissuer key tokenurl if not available from discovery URL specified by oidcissuer key attributesurl if not available from discovery URL specified by oidcissuer key jwksuri if not available from discovery URL specified by oidcissuer key For SAML providers: MetadataFile OR MetadataURL IDPSignout (boolean) optional

"identityProviderName": The name of the identity provider.

"identityProviderType": The type of the identity provider.

Main.Workspaces_Web.update_network_settings — Method

update_network_settings(network_settings_arn)
update_network_settings(network_settings_arn, params::Dict{String,<:Any})

Updates network settings.

Arguments

network_settings_arn: The ARN of the network settings.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, subsequent retries with the same client token return the result from the original successful request. If you do not specify a client token, one is automatically generated by the AWS SDK.

"securityGroupIds": One or more security groups used to control access from streaming instances to your VPC.

"subnetIds": The subnets in which network interfaces are created to connect streaming instances to your VPC. At least two of these subnets must be in different availability zones.

"vpcId": The VPC that streaming instances will connect to.

Main.Workspaces_Web.update_portal — Method

update_portal(portal_arn)
update_portal(portal_arn, params::Dict{String,<:Any})

Updates a web portal.

Arguments

portal_arn: The ARN of the web portal.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"authenticationType": The type of authentication integration points used when signing into the web portal. Defaults to Standard. Standard web portals are authenticated directly through your identity provider. You need to call CreateIdentityProvider to integrate your identity provider with your web portal. User and group access to your web portal is controlled through your identity provider. IAMIdentityCenter web portals are authenticated through AWS IAM Identity Center (successor to AWS Single Sign-On). They provide additional features, such as IdP-initiated authentication. Identity sources (including external identity provider integration), plus user and group access to your web portal, can be configured in the IAM Identity Center.

"displayName": The name of the web portal. This is not visible to users who log into the web portal.

Main.Workspaces_Web.update_trust_store — Method

update_trust_store(trust_store_arn)
update_trust_store(trust_store_arn, params::Dict{String,<:Any})

Updates the trust store.

Arguments

trust_store_arn: The ARN of the trust store.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"certificatesToAdd": A list of CA certificates to add to the trust store.

"certificatesToDelete": A list of CA certificates to delete from a trust store.

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, subsequent retries with the same client token return the result from the original successful request. If you do not specify a client token, one is automatically generated by the AWS SDK.

Main.Workspaces_Web.update_user_access_logging_settings — Method

update_user_access_logging_settings(user_access_logging_settings_arn)
update_user_access_logging_settings(user_access_logging_settings_arn, params::Dict{String,<:Any})

Updates the user access logging settings.

Arguments

user_access_logging_settings_arn: The ARN of the user access logging settings.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, subsequent retries with the same client token return the result from the original successful request. If you do not specify a client token, one is automatically generated by the AWS SDK.

"kinesisStreamArn": The ARN of the Kinesis stream.

Main.Workspaces_Web.update_user_settings — Method

update_user_settings(user_settings_arn)
update_user_settings(user_settings_arn, params::Dict{String,<:Any})

Updates the user settings.

Arguments

user_settings_arn: The ARN of the user settings.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request. Idempotency ensures that an API request completes only once. With an idempotent request, if the original request completes successfully, subsequent retries with the same client token return the result from the original successful request. If you do not specify a client token, one is automatically generated by the AWS SDK.

"copyAllowed": Specifies whether the user can copy text from the streaming session to the local device.

"disconnectTimeoutInMinutes": The amount of time that a streaming session remains active after users disconnect.

"downloadAllowed": Specifies whether the user can download files from the streaming session to the local device.

"idleDisconnectTimeoutInMinutes": The amount of time that users can be idle (inactive) before they are disconnected from their streaming session and the disconnect timeout interval begins.

"pasteAllowed": Specifies whether the user can paste text from the local device to the streaming session.

"printAllowed": Specifies whether the user can print to the local device.

"uploadAllowed": Specifies whether the user can upload files from the local device to the streaming session.