Appsync

This page documents function available when using the Appsync module, created with @service Appsync.

Index

Main.Appsync.create_api_cache

Main.Appsync.create_api_key

Main.Appsync.create_data_source

Main.Appsync.create_function

Main.Appsync.create_graphql_api

Main.Appsync.create_resolver

Main.Appsync.create_type

Main.Appsync.delete_api_cache

Main.Appsync.delete_api_key

Main.Appsync.delete_data_source

Main.Appsync.delete_function

Main.Appsync.delete_graphql_api

Main.Appsync.delete_resolver

Main.Appsync.delete_type

Main.Appsync.flush_api_cache

Main.Appsync.get_api_cache

Main.Appsync.get_data_source

Main.Appsync.get_function

Main.Appsync.get_graphql_api

Main.Appsync.get_introspection_schema

Main.Appsync.get_resolver

Main.Appsync.get_schema_creation_status

Main.Appsync.get_type

Main.Appsync.list_api_keys

Main.Appsync.list_data_sources

Main.Appsync.list_functions

Main.Appsync.list_graphql_apis

Main.Appsync.list_resolvers

Main.Appsync.list_resolvers_by_function

Main.Appsync.list_tags_for_resource

Main.Appsync.list_types

Main.Appsync.start_schema_creation

Main.Appsync.tag_resource

Main.Appsync.untag_resource

Main.Appsync.update_api_cache

Main.Appsync.update_api_key

Main.Appsync.update_data_source

Main.Appsync.update_function

Main.Appsync.update_graphql_api

Main.Appsync.update_resolver

Main.Appsync.update_type

Documentation

Main.Appsync.create_api_cache — Method

CreateApiCache()

Creates a cache for the GraphQL API.

Required Parameters

apiCachingBehavior: Caching behavior. FULLREQUESTCACHING: All requests are fully cached. PERRESOLVERCACHING: Individual resolvers that you specify are cached.

apiId: The GraphQL API Id.

ttl: TTL in seconds for cache entries. Valid values are between 1 and 3600 seconds.

type: The cache instance type. Valid values are SMALL MEDIUM LARGE XLARGE LARGE2X LARGE4X LARGE8X (not available in all regions) LARGE12X Historically, instance types were identified by an EC2-style value. As of July 2020, this is deprecated, and the generic identifiers above should be used. The following legacy instance types are available, but their use is discouraged: T2SMALL: A t2.small instance type. T2MEDIUM: A t2.medium instance type. R4LARGE: A r4.large instance type. R4XLARGE: A r4.xlarge instance type. R42XLARGE: A r4.2xlarge instance type. R44XLARGE: A r4.4xlarge instance type. R4_8XLARGE: A r4.8xlarge instance type.

Optional Parameters

atRestEncryptionEnabled: At rest encryption flag for cache. This setting cannot be updated after creation.

transitEncryptionEnabled: Transit encryption flag when connecting to cache. This setting cannot be updated after creation.

Main.Appsync.create_api_key — Method

CreateApiKey()

Creates a unique key that you can distribute to clients who are executing your API.

Required Parameters

apiId: The ID for your GraphQL API.

Optional Parameters

description: A description of the purpose of the API key.

expires: The time from creation time after which the API key expires. The date is represented as seconds since the epoch, rounded down to the nearest hour. The default value for this parameter is 7 days from creation time. For more information, see .

Main.Appsync.create_data_source — Method

CreateDataSource()

Creates a DataSource object.

Required Parameters

apiId: The API ID for the GraphQL API for the DataSource.

name: A user-supplied name for the DataSource.

type: The type of the DataSource.

Optional Parameters

description: A description of the DataSource.

dynamodbConfig: Amazon DynamoDB settings.

elasticsearchConfig: Amazon Elasticsearch Service settings.

httpConfig: HTTP endpoint settings.

lambdaConfig: AWS Lambda settings.

relationalDatabaseConfig: Relational database settings.

serviceRoleArn: The AWS IAM service role ARN for the data source. The system assumes this role when accessing the data source.

Main.Appsync.create_function — Method

CreateFunction()

Creates a Function object. A function is a reusable entity. Multiple functions can be used to compose the resolver logic.

Required Parameters

apiId: The GraphQL API ID.

dataSourceName: The Function DataSource name.

functionVersion: The version of the request mapping template. Currently the supported value is 2018-05-29.

name: The Function name. The function name does not have to be unique.

Optional Parameters

description: The Function description.

requestMappingTemplate: The Function request mapping template. Functions support only the 2018-05-29 version of the request mapping template.

responseMappingTemplate: The Function response mapping template.

Main.Appsync.create_graphql_api — Method

CreateGraphqlApi()

Creates a GraphqlApi object.

Required Parameters

authenticationType: The authentication type: API key, AWS IAM, OIDC, or Amazon Cognito user pools.

name: A user-supplied name for the GraphqlApi.

Optional Parameters

additionalAuthenticationProviders: A list of additional authentication providers for the GraphqlApi API.

logConfig: The Amazon CloudWatch Logs configuration.

openIDConnectConfig: The OpenID Connect configuration.

tags: A TagMap object.

userPoolConfig: The Amazon Cognito user pool configuration.

xrayEnabled: A flag indicating whether to enable X-Ray tracing for the GraphqlApi.

Main.Appsync.create_resolver — Method

CreateResolver()

Creates a Resolver object. A resolver converts incoming requests into a format that a data source can understand and converts the data source's responses into GraphQL.

Required Parameters

apiId: The ID for the GraphQL API for which the resolver is being created.

fieldName: The name of the field to attach the resolver to.

typeName: The name of the Type.

Optional Parameters

cachingConfig: The caching configuration for the resolver.

dataSourceName: The name of the data source for which the resolver is being created.

kind: The resolver type. UNIT: A UNIT resolver type. A UNIT resolver is the default resolver type. A UNIT resolver enables you to execute a GraphQL query against a single data source. PIPELINE: A PIPELINE resolver type. A PIPELINE resolver enables you to execute a series of Function in a serial manner. You can use a pipeline resolver to execute a GraphQL query against multiple data sources.

pipelineConfig: The PipelineConfig.

requestMappingTemplate: The mapping template to be used for requests. A resolver uses a request mapping template to convert a GraphQL expression into a format that a data source can understand. Mapping templates are written in Apache Velocity Template Language (VTL). VTL request mapping templates are optional when using a Lambda data source. For all other data sources, VTL request and response mapping templates are required.

responseMappingTemplate: The mapping template to be used for responses from the data source.

syncConfig: The SyncConfig for a resolver attached to a versioned datasource.

Main.Appsync.create_type — Method

CreateType()

Creates a Type object.

Required Parameters

apiId: The API ID.

definition: The type definition, in GraphQL Schema Definition Language (SDL) format. For more information, see the GraphQL SDL documentation.

format: The type format: SDL or JSON.

Main.Appsync.delete_api_cache — Method

DeleteApiCache()

Deletes an ApiCache object.

Required Parameters

apiId: The API ID.

Main.Appsync.delete_api_key — Method

DeleteApiKey()

Deletes an API key.

Required Parameters

apiId: The API ID.

id: The ID for the API key.

Main.Appsync.delete_data_source — Method

DeleteDataSource()

Deletes a DataSource object.

Required Parameters

apiId: The API ID.

name: The name of the data source.

Main.Appsync.delete_function — Method

DeleteFunction()

Deletes a Function.

Required Parameters

apiId: The GraphQL API ID.

functionId: The Function ID.

Main.Appsync.delete_graphql_api — Method

DeleteGraphqlApi()

Deletes a GraphqlApi object.

Required Parameters

apiId: The API ID.

Main.Appsync.delete_resolver — Method

DeleteResolver()

Deletes a Resolver object.

Required Parameters

apiId: The API ID.

fieldName: The resolver field name.

typeName: The name of the resolver type.

Main.Appsync.delete_type — Method

DeleteType()

Deletes a Type object.

Required Parameters

apiId: The API ID.

typeName: The type name.

Main.Appsync.flush_api_cache — Method

FlushApiCache()

Flushes an ApiCache object.

Required Parameters

apiId: The API ID.

Main.Appsync.get_api_cache — Method

GetApiCache()

Retrieves an ApiCache object.

Required Parameters

apiId: The API ID.

Main.Appsync.get_data_source — Method

GetDataSource()

Retrieves a DataSource object.

Required Parameters

apiId: The API ID.

name: The name of the data source.

Main.Appsync.get_function — Method

GetFunction()

Get a Function.

Required Parameters

apiId: The GraphQL API ID.

functionId: The Function ID.

Main.Appsync.get_graphql_api — Method

GetGraphqlApi()

Retrieves a GraphqlApi object.

Required Parameters

apiId: The API ID for the GraphQL API.

Main.Appsync.get_introspection_schema — Method

GetIntrospectionSchema()

Retrieves the introspection schema for a GraphQL API.

Required Parameters

apiId: The API ID.

format: The schema format: SDL or JSON.

Optional Parameters

includeDirectives: A flag that specifies whether the schema introspection should contain directives.

Main.Appsync.get_resolver — Method

GetResolver()

Retrieves a Resolver object.

Required Parameters

apiId: The API ID.

fieldName: The resolver field name.

typeName: The resolver type name.

Main.Appsync.get_schema_creation_status — Method

GetSchemaCreationStatus()

Retrieves the current status of a schema creation operation.

Required Parameters

apiId: The API ID.

Main.Appsync.get_type — Method

GetType()

Retrieves a Type object.

Required Parameters

apiId: The API ID.

format: The type format: SDL or JSON.

typeName: The type name.

Main.Appsync.list_api_keys — Method

ListApiKeys()

Lists the API keys for a given API. API keys are deleted automatically 60 days after they expire. However, they may still be included in the response until they have actually been deleted. You can safely call DeleteApiKey to manually delete a key before it's automatically deleted.

Required Parameters

apiId: The API ID.

Optional Parameters

maxResults: The maximum number of results you want the request to return.

nextToken: An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Main.Appsync.list_data_sources — Method

ListDataSources()

Lists the data sources for a given API.

Required Parameters

apiId: The API ID.

Optional Parameters

maxResults: The maximum number of results you want the request to return.

nextToken: An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Main.Appsync.list_functions — Method

ListFunctions()

List multiple functions.

Required Parameters

apiId: The GraphQL API ID.

Optional Parameters

maxResults: The maximum number of results you want the request to return.

nextToken: An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Main.Appsync.list_graphql_apis — Method

ListGraphqlApis()

Lists your GraphQL APIs.

Optional Parameters

maxResults: The maximum number of results you want the request to return.

nextToken: An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Main.Appsync.list_resolvers — Method

ListResolvers()

Lists the resolvers for a given API and type.

Required Parameters

apiId: The API ID.

typeName: The type name.

Optional Parameters

maxResults: The maximum number of results you want the request to return.

nextToken: An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Main.Appsync.list_resolvers_by_function — Method

ListResolversByFunction()

List the resolvers that are associated with a specific function.

Required Parameters

apiId: The API ID.

functionId: The Function ID.

Optional Parameters

maxResults: The maximum number of results you want the request to return.

nextToken: An identifier that was returned from the previous call to this operation, which you can use to return the next set of items in the list.

Main.Appsync.list_tags_for_resource — Method

ListTagsForResource()

Lists the tags for a resource.

Required Parameters

resourceArn: The GraphqlApi ARN.

Main.Appsync.list_types — Method

ListTypes()

Lists the types for a given API.

Required Parameters

apiId: The API ID.

format: The type format: SDL or JSON.

Optional Parameters

maxResults: The maximum number of results you want the request to return.

nextToken: An identifier that was returned from the previous call to this operation, which can be used to return the next set of items in the list.

Main.Appsync.start_schema_creation — Method

StartSchemaCreation()

Adds a new schema to your GraphQL API. This operation is asynchronous. Use to determine when it has completed.

Required Parameters

apiId: The API ID.

definition: The schema definition, in GraphQL schema language format.

Main.Appsync.tag_resource — Method

TagResource()

Tags a resource with user-supplied tags.

Required Parameters

resourceArn: The GraphqlApi ARN.

tags: A TagMap object.

Main.Appsync.untag_resource — Method

UntagResource()

Untags a resource.

Required Parameters

resourceArn: The GraphqlApi ARN.

tagKeys: A list of TagKey objects.

Main.Appsync.update_api_cache — Method

UpdateApiCache()

Updates the cache for the GraphQL API.

Required Parameters

apiCachingBehavior: Caching behavior. FULLREQUESTCACHING: All requests are fully cached. PERRESOLVERCACHING: Individual resolvers that you specify are cached.

apiId: The GraphQL API Id.

ttl: TTL in seconds for cache entries. Valid values are between 1 and 3600 seconds.

type: The cache instance type. Valid values are SMALL MEDIUM LARGE XLARGE LARGE2X LARGE4X LARGE8X (not available in all regions) LARGE12X Historically, instance types were identified by an EC2-style value. As of July 2020, this is deprecated, and the generic identifiers above should be used. The following legacy instance types are available, but their use is discouraged: T2SMALL: A t2.small instance type. T2MEDIUM: A t2.medium instance type. R4LARGE: A r4.large instance type. R4XLARGE: A r4.xlarge instance type. R42XLARGE: A r4.2xlarge instance type. R44XLARGE: A r4.4xlarge instance type. R4_8XLARGE: A r4.8xlarge instance type.

Main.Appsync.update_api_key — Method

UpdateApiKey()

Updates an API key. The key can be updated while it is not deleted.

Required Parameters

apiId: The ID for the GraphQL API.

id: The API key ID.

Optional Parameters

description: A description of the purpose of the API key.

expires: The time from update time after which the API key expires. The date is represented as seconds since the epoch. For more information, see .

Main.Appsync.update_data_source — Method

UpdateDataSource()

Updates a DataSource object.

Required Parameters

apiId: The API ID.

name: The new name for the data source.

type: The new data source type.

Optional Parameters

description: The new description for the data source.

dynamodbConfig: The new Amazon DynamoDB configuration.

elasticsearchConfig: The new Elasticsearch Service configuration.

httpConfig: The new HTTP endpoint configuration.

lambdaConfig: The new AWS Lambda configuration.

relationalDatabaseConfig: The new relational database configuration.

serviceRoleArn: The new service role ARN for the data source.

Main.Appsync.update_function — Method

UpdateFunction()

Updates a Function object.

Required Parameters

apiId: The GraphQL API ID.

dataSourceName: The Function DataSource name.

functionId: The function ID.

functionVersion: The version of the request mapping template. Currently the supported value is 2018-05-29.

name: The Function name.

Optional Parameters

description: The Function description.

requestMappingTemplate: The Function request mapping template. Functions support only the 2018-05-29 version of the request mapping template.

responseMappingTemplate: The Function request mapping template.

Main.Appsync.update_graphql_api — Method

UpdateGraphqlApi()

Updates a GraphqlApi object.

Required Parameters

apiId: The API ID.

name: The new name for the GraphqlApi object.

Optional Parameters

additionalAuthenticationProviders: A list of additional authentication providers for the GraphqlApi API.

authenticationType: The new authentication type for the GraphqlApi object.

logConfig: The Amazon CloudWatch Logs configuration for the GraphqlApi object.

openIDConnectConfig: The OpenID Connect configuration for the GraphqlApi object.

userPoolConfig: The new Amazon Cognito user pool configuration for the GraphqlApi object.

xrayEnabled: A flag indicating whether to enable X-Ray tracing for the GraphqlApi.

Main.Appsync.update_resolver — Method

UpdateResolver()

Updates a Resolver object.

Required Parameters

apiId: The API ID.

fieldName: The new field name.

typeName: The new type name.

Optional Parameters

cachingConfig: The caching configuration for the resolver.

dataSourceName: The new data source name.

kind: The resolver type. UNIT: A UNIT resolver type. A UNIT resolver is the default resolver type. A UNIT resolver enables you to execute a GraphQL query against a single data source. PIPELINE: A PIPELINE resolver type. A PIPELINE resolver enables you to execute a series of Function in a serial manner. You can use a pipeline resolver to execute a GraphQL query against multiple data sources.

pipelineConfig: The PipelineConfig.

requestMappingTemplate: The new request mapping template. A resolver uses a request mapping template to convert a GraphQL expression into a format that a data source can understand. Mapping templates are written in Apache Velocity Template Language (VTL). VTL request mapping templates are optional when using a Lambda data source. For all other data sources, VTL request and response mapping templates are required.

responseMappingTemplate: The new response mapping template.

syncConfig: The SyncConfig for a resolver attached to a versioned datasource.

Main.Appsync.update_type — Method

UpdateType()

Updates a Type object.

Required Parameters

apiId: The API ID.

format: The new type format: SDL or JSON.

typeName: The new type name.

Optional Parameters

definition: The new definition.