Appsync

This page documents function available when using the Appsync module, created with @service Appsync.

Index

Main.Appsync.associate_api

Main.Appsync.create_api_cache

Main.Appsync.create_api_key

Main.Appsync.create_data_source

Main.Appsync.create_domain_name

Main.Appsync.create_function

Main.Appsync.create_graphql_api

Main.Appsync.create_resolver

Main.Appsync.create_type

Main.Appsync.delete_api_cache

Main.Appsync.delete_api_key

Main.Appsync.delete_data_source

Main.Appsync.delete_domain_name

Main.Appsync.delete_function

Main.Appsync.delete_graphql_api

Main.Appsync.delete_resolver

Main.Appsync.delete_type

Main.Appsync.disassociate_api

Main.Appsync.evaluate_code

Main.Appsync.evaluate_mapping_template

Main.Appsync.flush_api_cache

Main.Appsync.get_api_association

Main.Appsync.get_api_cache

Main.Appsync.get_data_source

Main.Appsync.get_domain_name

Main.Appsync.get_function

Main.Appsync.get_graphql_api

Main.Appsync.get_introspection_schema

Main.Appsync.get_resolver

Main.Appsync.get_schema_creation_status

Main.Appsync.get_type

Main.Appsync.list_api_keys

Main.Appsync.list_data_sources

Main.Appsync.list_domain_names

Main.Appsync.list_functions

Main.Appsync.list_graphql_apis

Main.Appsync.list_resolvers

Main.Appsync.list_resolvers_by_function

Main.Appsync.list_tags_for_resource

Main.Appsync.list_types

Main.Appsync.start_schema_creation

Main.Appsync.tag_resource

Main.Appsync.untag_resource

Main.Appsync.update_api_cache

Main.Appsync.update_api_key

Main.Appsync.update_data_source

Main.Appsync.update_domain_name

Main.Appsync.update_function

Main.Appsync.update_graphql_api

Main.Appsync.update_resolver

Main.Appsync.update_type

Documentation

Main.Appsync.associate_api — Method

associate_api(api_id, domain_name)
associate_api(api_id, domain_name, params::Dict{String,<:Any})

Maps an endpoint to your custom domain.

Arguments

api_id: The API ID.

domain_name: The domain name.

Main.Appsync.create_api_cache — Method

create_api_cache(api_caching_behavior, api_id, ttl, type)
create_api_cache(api_caching_behavior, api_id, ttl, type, params::Dict{String,<:Any})

Creates a cache for the GraphQL API.

Arguments

api_caching_behavior: Caching behavior. FULLREQUESTCACHING: All requests are fully cached. PERRESOLVERCACHING: Individual resolvers that you specify are cached.

api_id: The GraphQL API ID.

ttl: TTL in seconds for cache entries. Valid values are 1–3,600 seconds.

type: The cache instance type. Valid values are SMALL MEDIUM LARGE XLARGE LARGE2X LARGE4X LARGE8X (not available in all regions) LARGE12X Historically, instance types were identified by an EC2-style value. As of July 2020, this is deprecated, and the generic identifiers above should be used. The following legacy instance types are available, but their use is discouraged: T2SMALL: A t2.small instance type. T2MEDIUM: A t2.medium instance type. R4LARGE: A r4.large instance type. R4XLARGE: A r4.xlarge instance type. R42XLARGE: A r4.2xlarge instance type. R44XLARGE: A r4.4xlarge instance type. R4_8XLARGE: A r4.8xlarge instance type.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"atRestEncryptionEnabled": At-rest encryption flag for cache. You cannot update this setting after creation.

"transitEncryptionEnabled": Transit encryption flag when connecting to cache. You cannot update this setting after creation.

Main.Appsync.create_api_key — Method

create_api_key(api_id)
create_api_key(api_id, params::Dict{String,<:Any})

Creates a unique key that you can distribute to clients who invoke your API.

Arguments

api_id: The ID for your GraphQL API.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A description of the purpose of the API key.

"expires": From the creation time, the time after which the API key expires. The date is represented as seconds since the epoch, rounded down to the nearest hour. The default value for this parameter is 7 days from creation time. For more information, see .

Main.Appsync.create_data_source — Method

create_data_source(api_id, name, type)
create_data_source(api_id, name, type, params::Dict{String,<:Any})

Creates a DataSource object.

Arguments

api_id: The API ID for the GraphQL API for the DataSource.

name: A user-supplied name for the DataSource.

type: The type of the DataSource.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A description of the DataSource.

"dynamodbConfig": Amazon DynamoDB settings.

"elasticsearchConfig": Amazon OpenSearch Service settings. As of September 2021, Amazon Elasticsearch service is Amazon OpenSearch Service. This configuration is deprecated. For new data sources, use CreateDataSourceRequestopenSearchServiceConfig to create an OpenSearch data source.

"eventBridgeConfig": Amazon EventBridge settings.

"httpConfig": HTTP endpoint settings.

"lambdaConfig": Lambda settings.

"openSearchServiceConfig": Amazon OpenSearch Service settings.

"relationalDatabaseConfig": Relational database settings.

"serviceRoleArn": The Identity and Access Management (IAM) service role Amazon Resource Name (ARN) for the data source. The system assumes this role when accessing the data source.

Main.Appsync.create_domain_name — Method

create_domain_name(certificate_arn, domain_name)
create_domain_name(certificate_arn, domain_name, params::Dict{String,<:Any})

Creates a custom DomainName object.

Arguments

certificate_arn: The Amazon Resource Name (ARN) of the certificate. This can be an Certificate Manager (ACM) certificate or an Identity and Access Management (IAM) server certificate.

domain_name: The domain name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A description of the DomainName.

Main.Appsync.create_function — Method

create_function(api_id, data_source_name, name)
create_function(api_id, data_source_name, name, params::Dict{String,<:Any})

Creates a Function object. A function is a reusable entity. You can use multiple functions to compose the resolver logic.

Arguments

api_id: The GraphQL API ID.

data_source_name: The Function DataSource name.

name: The Function name. The function name does not have to be unique.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"code": The function code that contains the request and response functions. When code is used, the runtime is required. The runtime value must be APPSYNC_JS.

"description": The Function description.

"functionVersion": The version of the request mapping template. Currently, the supported value is 2018-05-29. Note that when using VTL and mapping templates, the functionVersion is required.

"maxBatchSize": The maximum batching size for a resolver.

"requestMappingTemplate": The Function request mapping template. Functions support only the 2018-05-29 version of the request mapping template.

"responseMappingTemplate": The Function response mapping template.

"runtime":

"syncConfig":

Main.Appsync.create_graphql_api — Method

create_graphql_api(authentication_type, name)
create_graphql_api(authentication_type, name, params::Dict{String,<:Any})

Creates a GraphqlApi object.

Arguments

authentication_type: The authentication type: API key, Identity and Access Management (IAM), OpenID Connect (OIDC), Amazon Cognito user pools, or Lambda.

name: A user-supplied name for the GraphqlApi.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"additionalAuthenticationProviders": A list of additional authentication providers for the GraphqlApi API.

"lambdaAuthorizerConfig": Configuration for Lambda function authorization.

"logConfig": The Amazon CloudWatch Logs configuration.

"openIDConnectConfig": The OIDC configuration.

"tags": A TagMap object.

"userPoolConfig": The Amazon Cognito user pool configuration.

"xrayEnabled": A flag indicating whether to use X-Ray tracing for the GraphqlApi.

Main.Appsync.create_resolver — Method

create_resolver(api_id, field_name, type_name)
create_resolver(api_id, field_name, type_name, params::Dict{String,<:Any})

Creates a Resolver object. A resolver converts incoming requests into a format that a data source can understand, and converts the data source's responses into GraphQL.

Arguments

api_id: The ID for the GraphQL API for which the resolver is being created.

field_name: The name of the field to attach the resolver to.

type_name: The name of the Type.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"cachingConfig": The caching configuration for the resolver.

"code": The resolver code that contains the request and response functions. When code is used, the runtime is required. The runtime value must be APPSYNC_JS.

"dataSourceName": The name of the data source for which the resolver is being created.

"kind": The resolver type. UNIT: A UNIT resolver type. A UNIT resolver is the default resolver type. You can use a UNIT resolver to run a GraphQL query against a single data source. PIPELINE: A PIPELINE resolver type. You can use a PIPELINE resolver to invoke a series of Function objects in a serial manner. You can use a pipeline resolver to run a GraphQL query against multiple data sources.

"maxBatchSize": The maximum batching size for a resolver.

"pipelineConfig": The PipelineConfig.

"requestMappingTemplate": The mapping template to use for requests. A resolver uses a request mapping template to convert a GraphQL expression into a format that a data source can understand. Mapping templates are written in Apache Velocity Template Language (VTL). VTL request mapping templates are optional when using an Lambda data source. For all other data sources, VTL request and response mapping templates are required.

"responseMappingTemplate": The mapping template to use for responses from the data source.

"runtime":

"syncConfig": The SyncConfig for a resolver attached to a versioned data source.

Main.Appsync.create_type — Method

create_type(api_id, definition, format)
create_type(api_id, definition, format, params::Dict{String,<:Any})

Creates a Type object.

Arguments

api_id: The API ID.

definition: The type definition, in GraphQL Schema Definition Language (SDL) format. For more information, see the GraphQL SDL documentation.

format: The type format: SDL or JSON.

Main.Appsync.delete_api_cache — Method

delete_api_cache(api_id)
delete_api_cache(api_id, params::Dict{String,<:Any})

Deletes an ApiCache object.

Arguments

api_id: The API ID.

Main.Appsync.delete_api_key — Method

delete_api_key(api_id, id)
delete_api_key(api_id, id, params::Dict{String,<:Any})

Deletes an API key.

Arguments

api_id: The API ID.

id: The ID for the API key.

Main.Appsync.delete_data_source — Method

delete_data_source(api_id, name)
delete_data_source(api_id, name, params::Dict{String,<:Any})

Deletes a DataSource object.

Arguments

api_id: The API ID.

name: The name of the data source.

Main.Appsync.delete_domain_name — Method

delete_domain_name(domain_name)
delete_domain_name(domain_name, params::Dict{String,<:Any})

Deletes a custom DomainName object.

Arguments

domain_name: The domain name.

Main.Appsync.delete_function — Method

delete_function(api_id, function_id)
delete_function(api_id, function_id, params::Dict{String,<:Any})

Deletes a Function.

Arguments

api_id: The GraphQL API ID.

function_id: The Function ID.

Main.Appsync.delete_graphql_api — Method

delete_graphql_api(api_id)
delete_graphql_api(api_id, params::Dict{String,<:Any})

Deletes a GraphqlApi object.

Arguments

api_id: The API ID.

Main.Appsync.delete_resolver — Method

delete_resolver(api_id, field_name, type_name)
delete_resolver(api_id, field_name, type_name, params::Dict{String,<:Any})

Deletes a Resolver object.

Arguments

api_id: The API ID.

field_name: The resolver field name.

type_name: The name of the resolver type.

Main.Appsync.delete_type — Method

delete_type(api_id, type_name)
delete_type(api_id, type_name, params::Dict{String,<:Any})

Deletes a Type object.

Arguments

api_id: The API ID.

type_name: The type name.

Main.Appsync.disassociate_api — Method

disassociate_api(domain_name)
disassociate_api(domain_name, params::Dict{String,<:Any})

Removes an ApiAssociation object from a custom domain.

Arguments

domain_name: The domain name.

Main.Appsync.evaluate_code — Method

evaluate_code(code, context, runtime)
evaluate_code(code, context, runtime, params::Dict{String,<:Any})

Evaluates the given code and returns the response. The code definition requirements depend on the specified runtime. For APPSYNC_JS runtimes, the code defines the request and response functions. The request function takes the incoming request after a GraphQL operation is parsed and converts it into a request configuration for the selected data source operation. The response function interprets responses from the data source and maps it to the shape of the GraphQL field output type.

Arguments

code: The code definition to be evaluated. Note that code and runtime are both required for this action. The runtime value must be APPSYNC_JS.

context: The map that holds all of the contextual information for your resolver invocation. A context is required for this action.

runtime: The runtime to be used when evaluating the code. Currently, only the APPSYNC_JS runtime is supported.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"function": The function within the code to be evaluated. If provided, the valid values are request and response.

Main.Appsync.evaluate_mapping_template — Method

evaluate_mapping_template(context, template)
evaluate_mapping_template(context, template, params::Dict{String,<:Any})

Evaluates a given template and returns the response. The mapping template can be a request or response template. Request templates take the incoming request after a GraphQL operation is parsed and convert it into a request configuration for the selected data source operation. Response templates interpret responses from the data source and map it to the shape of the GraphQL field output type. Mapping templates are written in the Apache Velocity Template Language (VTL).

Arguments

context: The map that holds all of the contextual information for your resolver invocation. A context is required for this action.

template: The mapping template; this can be a request or response template. A template is required for this action.

Main.Appsync.flush_api_cache — Method

flush_api_cache(api_id)
flush_api_cache(api_id, params::Dict{String,<:Any})

Flushes an ApiCache object.

Arguments

api_id: The API ID.

Main.Appsync.get_api_association — Method

get_api_association(domain_name)
get_api_association(domain_name, params::Dict{String,<:Any})

Retrieves an ApiAssociation object.

Arguments

domain_name: The domain name.

Main.Appsync.get_api_cache — Method

get_api_cache(api_id)
get_api_cache(api_id, params::Dict{String,<:Any})

Retrieves an ApiCache object.

Arguments

api_id: The API ID.

Main.Appsync.get_data_source — Method

get_data_source(api_id, name)
get_data_source(api_id, name, params::Dict{String,<:Any})

Retrieves a DataSource object.

Arguments

api_id: The API ID.

name: The name of the data source.

Main.Appsync.get_domain_name — Method

get_domain_name(domain_name)
get_domain_name(domain_name, params::Dict{String,<:Any})

Retrieves a custom DomainName object.

Arguments

domain_name: The domain name.

Main.Appsync.get_function — Method

get_function(api_id, function_id)
get_function(api_id, function_id, params::Dict{String,<:Any})

Get a Function.

Arguments

api_id: The GraphQL API ID.

function_id: The Function ID.

Main.Appsync.get_graphql_api — Method

get_graphql_api(api_id)
get_graphql_api(api_id, params::Dict{String,<:Any})

Retrieves a GraphqlApi object.

Arguments

api_id: The API ID for the GraphQL API.

Main.Appsync.get_introspection_schema — Method

get_introspection_schema(api_id, format)
get_introspection_schema(api_id, format, params::Dict{String,<:Any})

Retrieves the introspection schema for a GraphQL API.

Arguments

api_id: The API ID.

format: The schema format: SDL or JSON.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"includeDirectives": A flag that specifies whether the schema introspection should contain directives.

Main.Appsync.get_resolver — Method

get_resolver(api_id, field_name, type_name)
get_resolver(api_id, field_name, type_name, params::Dict{String,<:Any})

Retrieves a Resolver object.

Arguments

api_id: The API ID.

field_name: The resolver field name.

type_name: The resolver type name.

Main.Appsync.get_schema_creation_status — Method

get_schema_creation_status(api_id)
get_schema_creation_status(api_id, params::Dict{String,<:Any})

Retrieves the current status of a schema creation operation.

Arguments

api_id: The API ID.

Main.Appsync.get_type — Method

get_type(api_id, format, type_name)
get_type(api_id, format, type_name, params::Dict{String,<:Any})

Retrieves a Type object.

Arguments

api_id: The API ID.

format: The type format: SDL or JSON.

type_name: The type name.

Main.Appsync.list_api_keys — Method

list_api_keys(api_id)
list_api_keys(api_id, params::Dict{String,<:Any})

Lists the API keys for a given API. API keys are deleted automatically 60 days after they expire. However, they may still be included in the response until they have actually been deleted. You can safely call DeleteApiKey to manually delete a key before it's automatically deleted.

Arguments

api_id: The API ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results that you want the request to return.

"nextToken": An identifier that was returned from the previous call to this operation, which you can use to return the next set of items in the list.

Main.Appsync.list_data_sources — Method

list_data_sources(api_id)
list_data_sources(api_id, params::Dict{String,<:Any})

Lists the data sources for a given API.

Arguments

api_id: The API ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results that you want the request to return.

"nextToken": An identifier that was returned from the previous call to this operation, which you can use to return the next set of items in the list.

Main.Appsync.list_domain_names — Method

list_domain_names()
list_domain_names(params::Dict{String,<:Any})

Lists multiple custom domain names.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results that you want the request to return.

"nextToken": The API token.

Main.Appsync.list_functions — Method

list_functions(api_id)
list_functions(api_id, params::Dict{String,<:Any})

List multiple functions.

Arguments

api_id: The GraphQL API ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results that you want the request to return.

"nextToken": An identifier that was returned from the previous call to this operation, which you can use to return the next set of items in the list.

Main.Appsync.list_graphql_apis — Method

list_graphql_apis()
list_graphql_apis(params::Dict{String,<:Any})

Lists your GraphQL APIs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results that you want the request to return.

"nextToken": An identifier that was returned from the previous call to this operation, which you can use to return the next set of items in the list.

Main.Appsync.list_resolvers — Method

list_resolvers(api_id, type_name)
list_resolvers(api_id, type_name, params::Dict{String,<:Any})

Lists the resolvers for a given API and type.

Arguments

api_id: The API ID.

type_name: The type name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results that you want the request to return.

"nextToken": An identifier that was returned from the previous call to this operation, which you can use to return the next set of items in the list.

Main.Appsync.list_resolvers_by_function — Method

list_resolvers_by_function(api_id, function_id)
list_resolvers_by_function(api_id, function_id, params::Dict{String,<:Any})

List the resolvers that are associated with a specific function.

Arguments

api_id: The API ID.

function_id: The function ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results that you want the request to return.

"nextToken": An identifier that was returned from the previous call to this operation, which you can use to return the next set of items in the list.

Main.Appsync.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists the tags for a resource.

Arguments

resource_arn: The GraphqlApi Amazon Resource Name (ARN).

Main.Appsync.list_types — Method

list_types(api_id, format)
list_types(api_id, format, params::Dict{String,<:Any})

Lists the types for a given API.

Arguments

api_id: The API ID.

format: The type format: SDL or JSON.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results that you want the request to return.

"nextToken": An identifier that was returned from the previous call to this operation, which you can use to return the next set of items in the list.

Main.Appsync.start_schema_creation — Method

start_schema_creation(api_id, definition)
start_schema_creation(api_id, definition, params::Dict{String,<:Any})

Adds a new schema to your GraphQL API. This operation is asynchronous. Use to determine when it has completed.

Arguments

api_id: The API ID.

definition: The schema definition, in GraphQL schema language format.

Main.Appsync.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Tags a resource with user-supplied tags.

Arguments

resource_arn: The GraphqlApi Amazon Resource Name (ARN).

tags: A TagMap object.

Main.Appsync.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Untags a resource.

Arguments

resource_arn: The GraphqlApi Amazon Resource Name (ARN).

tag_keys: A list of TagKey objects.

Main.Appsync.update_api_cache — Method

update_api_cache(api_caching_behavior, api_id, ttl, type)
update_api_cache(api_caching_behavior, api_id, ttl, type, params::Dict{String,<:Any})

Updates the cache for the GraphQL API.

Arguments

api_caching_behavior: Caching behavior. FULLREQUESTCACHING: All requests are fully cached. PERRESOLVERCACHING: Individual resolvers that you specify are cached.

api_id: The GraphQL API ID.

ttl: TTL in seconds for cache entries. Valid values are 1–3,600 seconds.

type: The cache instance type. Valid values are SMALL MEDIUM LARGE XLARGE LARGE2X LARGE4X LARGE8X (not available in all regions) LARGE12X Historically, instance types were identified by an EC2-style value. As of July 2020, this is deprecated, and the generic identifiers above should be used. The following legacy instance types are available, but their use is discouraged: T2SMALL: A t2.small instance type. T2MEDIUM: A t2.medium instance type. R4LARGE: A r4.large instance type. R4XLARGE: A r4.xlarge instance type. R42XLARGE: A r4.2xlarge instance type. R44XLARGE: A r4.4xlarge instance type. R4_8XLARGE: A r4.8xlarge instance type.

Main.Appsync.update_api_key — Method

update_api_key(api_id, id)
update_api_key(api_id, id, params::Dict{String,<:Any})

Updates an API key. You can update the key as long as it's not deleted.

Arguments

api_id: The ID for the GraphQL API.

id: The API key ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A description of the purpose of the API key.

"expires": From the update time, the time after which the API key expires. The date is represented as seconds since the epoch. For more information, see .

Main.Appsync.update_data_source — Method

update_data_source(api_id, name, type)
update_data_source(api_id, name, type, params::Dict{String,<:Any})

Updates a DataSource object.

Arguments

api_id: The API ID.

name: The new name for the data source.

type: The new data source type.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The new description for the data source.

"dynamodbConfig": The new Amazon DynamoDB configuration.

"elasticsearchConfig": The new OpenSearch configuration. As of September 2021, Amazon Elasticsearch service is Amazon OpenSearch Service. This configuration is deprecated. Instead, use UpdateDataSourceRequestopenSearchServiceConfig to update an OpenSearch data source.

"eventBridgeConfig": The new Amazon EventBridge settings.

"httpConfig": The new HTTP endpoint configuration.

"lambdaConfig": The new Lambda configuration.

"openSearchServiceConfig": The new OpenSearch configuration.

"relationalDatabaseConfig": The new relational database configuration.

"serviceRoleArn": The new service role Amazon Resource Name (ARN) for the data source.

Main.Appsync.update_domain_name — Method

update_domain_name(domain_name)
update_domain_name(domain_name, params::Dict{String,<:Any})

Updates a custom DomainName object.

Arguments

domain_name: The domain name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A description of the DomainName.

Main.Appsync.update_function — Method

update_function(api_id, data_source_name, function_id, name)
update_function(api_id, data_source_name, function_id, name, params::Dict{String,<:Any})

Updates a Function object.

Arguments

api_id: The GraphQL API ID.

data_source_name: The Function DataSource name.

function_id: The function ID.

name: The Function name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"code": The function code that contains the request and response functions. When code is used, the runtime is required. The runtime value must be APPSYNC_JS.

"description": The Function description.

"functionVersion": The version of the request mapping template. Currently, the supported value is 2018-05-29. Note that when using VTL and mapping templates, the functionVersion is required.

"maxBatchSize": The maximum batching size for a resolver.

"requestMappingTemplate": The Function request mapping template. Functions support only the 2018-05-29 version of the request mapping template.

"responseMappingTemplate": The Function request mapping template.

"runtime":

"syncConfig":

Main.Appsync.update_graphql_api — Method

update_graphql_api(api_id, name)
update_graphql_api(api_id, name, params::Dict{String,<:Any})

Updates a GraphqlApi object.

Arguments

api_id: The API ID.

name: The new name for the GraphqlApi object.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"additionalAuthenticationProviders": A list of additional authentication providers for the GraphqlApi API.

"authenticationType": The new authentication type for the GraphqlApi object.

"lambdaAuthorizerConfig": Configuration for Lambda function authorization.

"logConfig": The Amazon CloudWatch Logs configuration for the GraphqlApi object.

"openIDConnectConfig": The OpenID Connect configuration for the GraphqlApi object.

"userPoolConfig": The new Amazon Cognito user pool configuration for the ~GraphqlApi object.

"xrayEnabled": A flag indicating whether to use X-Ray tracing for the GraphqlApi.

Main.Appsync.update_resolver — Method

update_resolver(api_id, field_name, type_name)
update_resolver(api_id, field_name, type_name, params::Dict{String,<:Any})

Updates a Resolver object.

Arguments

api_id: The API ID.

field_name: The new field name.

type_name: The new type name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"cachingConfig": The caching configuration for the resolver.

"code": The resolver code that contains the request and response functions. When code is used, the runtime is required. The runtime value must be APPSYNC_JS.

"dataSourceName": The new data source name.

"kind": The resolver type. UNIT: A UNIT resolver type. A UNIT resolver is the default resolver type. You can use a UNIT resolver to run a GraphQL query against a single data source. PIPELINE: A PIPELINE resolver type. You can use a PIPELINE resolver to invoke a series of Function objects in a serial manner. You can use a pipeline resolver to run a GraphQL query against multiple data sources.

"maxBatchSize": The maximum batching size for a resolver.

"pipelineConfig": The PipelineConfig.

"requestMappingTemplate": The new request mapping template. A resolver uses a request mapping template to convert a GraphQL expression into a format that a data source can understand. Mapping templates are written in Apache Velocity Template Language (VTL). VTL request mapping templates are optional when using an Lambda data source. For all other data sources, VTL request and response mapping templates are required.

"responseMappingTemplate": The new response mapping template.

"runtime":

"syncConfig": The SyncConfig for a resolver attached to a versioned data source.

Main.Appsync.update_type — Method

update_type(api_id, format, type_name)
update_type(api_id, format, type_name, params::Dict{String,<:Any})

Updates a Type object.

Arguments

api_id: The API ID.

format: The new type format: SDL or JSON.

type_name: The new type name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"definition": The new definition.