Chatbot

This page documents function available when using the Chatbot module, created with @service Chatbot.

Index

Main.Chatbot.create_chime_webhook_configuration

Main.Chatbot.create_microsoft_teams_channel_configuration

Main.Chatbot.create_slack_channel_configuration

Main.Chatbot.delete_chime_webhook_configuration

Main.Chatbot.delete_microsoft_teams_channel_configuration

Main.Chatbot.delete_microsoft_teams_configured_team

Main.Chatbot.delete_microsoft_teams_user_identity

Main.Chatbot.delete_slack_channel_configuration

Main.Chatbot.delete_slack_user_identity

Main.Chatbot.delete_slack_workspace_authorization

Main.Chatbot.describe_chime_webhook_configurations

Main.Chatbot.describe_slack_channel_configurations

Main.Chatbot.describe_slack_user_identities

Main.Chatbot.describe_slack_workspaces

Main.Chatbot.get_account_preferences

Main.Chatbot.get_microsoft_teams_channel_configuration

Main.Chatbot.list_microsoft_teams_channel_configurations

Main.Chatbot.list_microsoft_teams_configured_teams

Main.Chatbot.list_microsoft_teams_user_identities

Main.Chatbot.list_tags_for_resource

Main.Chatbot.tag_resource

Main.Chatbot.untag_resource

Main.Chatbot.update_account_preferences

Main.Chatbot.update_chime_webhook_configuration

Main.Chatbot.update_microsoft_teams_channel_configuration

Main.Chatbot.update_slack_channel_configuration

Documentation

Main.Chatbot.create_chime_webhook_configuration — Method

create_chime_webhook_configuration(configuration_name, iam_role_arn, sns_topic_arns, webhook_description, webhook_url)
create_chime_webhook_configuration(configuration_name, iam_role_arn, sns_topic_arns, webhook_description, webhook_url, params::Dict{String,<:Any})

Creates Chime Webhook Configuration

Arguments

configuration_name: The name of the configuration.

iam_role_arn: This is a user-defined role that AWS Chatbot will assume. This is not the service-linked role. For more information, see IAM Policies for AWS Chatbot.

sns_topic_arns: The ARNs of the SNS topics that deliver notifications to AWS Chatbot.

webhook_description: Description of the webhook. Recommend using the convention RoomName/WebhookName. See Chime setup tutorial for more details: https://docs.aws.amazon.com/chatbot/latest/adminguide/chime-setup.html.

webhook_url: URL for the Chime webhook.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"LoggingLevel": Logging levels include ERROR, INFO, or NONE.

"Tags": A list of tags to apply to the configuration.

Main.Chatbot.create_microsoft_teams_channel_configuration — Method

create_microsoft_teams_channel_configuration(channel_id, configuration_name, iam_role_arn, team_id, tenant_id)
create_microsoft_teams_channel_configuration(channel_id, configuration_name, iam_role_arn, team_id, tenant_id, params::Dict{String,<:Any})

Creates MS Teams Channel Configuration

Arguments

channel_id: The ID of the Microsoft Teams channel.

configuration_name: The name of the configuration.

iam_role_arn: The ARN of the IAM role that defines the permissions for AWS Chatbot. This is a user-defined role that AWS Chatbot will assume. This is not the service-linked role. For more information, see IAM Policies for AWS Chatbot.

team_id: The ID of the Microsoft Team authorized with AWS Chatbot. To get the team ID, you must perform the initial authorization flow with Microsoft Teams in the AWS Chatbot console. Then you can copy and paste the team ID from the console. For more details, see steps 1-4 in Get started with Microsoft Teams in the AWS Chatbot Administrator Guide.

tenant_id: The ID of the Microsoft Teams tenant.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ChannelName": The name of the Microsoft Teams channel.

"GuardrailPolicyArns": The list of IAM policy ARNs that are applied as channel guardrails. The AWS managed 'AdministratorAccess' policy is applied by default if this is not set.

"LoggingLevel": Logging levels include ERROR, INFO, or NONE.

"SnsTopicArns": The ARNs of the SNS topics that deliver notifications to AWS Chatbot.

"Tags": A list of tags to apply to the configuration.

"TeamName": The name of the Microsoft Teams Team.

"UserAuthorizationRequired": Enables use of a user role requirement in your chat configuration.

Main.Chatbot.create_slack_channel_configuration — Method

create_slack_channel_configuration(configuration_name, iam_role_arn, slack_channel_id, slack_team_id)
create_slack_channel_configuration(configuration_name, iam_role_arn, slack_channel_id, slack_team_id, params::Dict{String,<:Any})

Creates Slack Channel Configuration

Arguments

configuration_name: The name of the configuration.

iam_role_arn: The ARN of the IAM role that defines the permissions for AWS Chatbot. This is a user-defined role that AWS Chatbot will assume. This is not the service-linked role. For more information, see IAM Policies for AWS Chatbot.

slack_channel_id: The ID of the Slack channel. To get the ID, open Slack, right click on the channel name in the left pane, then choose Copy Link. The channel ID is the 9-character string at the end of the URL. For example, ABCBBLZZZ.

slack_team_id: The ID of the Slack workspace authorized with AWS Chatbot.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"GuardrailPolicyArns": The list of IAM policy ARNs that are applied as channel guardrails. The AWS managed 'AdministratorAccess' policy is applied by default if this is not set.

"LoggingLevel": Logging levels include ERROR, INFO, or NONE.

"SlackChannelName": The name of the Slack Channel.

"SnsTopicArns": The ARNs of the SNS topics that deliver notifications to AWS Chatbot.

"Tags": A list of tags to apply to the configuration.

"UserAuthorizationRequired": Enables use of a user role requirement in your chat configuration.

Main.Chatbot.delete_chime_webhook_configuration — Method

delete_chime_webhook_configuration(chat_configuration_arn)
delete_chime_webhook_configuration(chat_configuration_arn, params::Dict{String,<:Any})

Deletes a Chime Webhook Configuration

Arguments

chat_configuration_arn: The ARN of the ChimeWebhookConfiguration to delete.

Main.Chatbot.delete_microsoft_teams_channel_configuration — Method

delete_microsoft_teams_channel_configuration(chat_configuration_arn)
delete_microsoft_teams_channel_configuration(chat_configuration_arn, params::Dict{String,<:Any})

Deletes MS Teams Channel Configuration

Arguments

chat_configuration_arn: The ARN of the MicrosoftTeamsChannelConfiguration to delete.

Main.Chatbot.delete_microsoft_teams_configured_team — Method

delete_microsoft_teams_configured_team(team_id)
delete_microsoft_teams_configured_team(team_id, params::Dict{String,<:Any})

Deletes the Microsoft Teams team authorization allowing for channels to be configured in that Microsoft Teams team. Note that the Microsoft Teams team must have no channels configured to remove it.

Arguments

team_id: The ID of the Microsoft Team authorized with AWS Chatbot. To get the team ID, you must perform the initial authorization flow with Microsoft Teams in the AWS Chatbot console. Then you can copy and paste the team ID from the console. For more details, see steps 1-4 in Get started with Microsoft Teams in the AWS Chatbot Administrator Guide.

Main.Chatbot.delete_microsoft_teams_user_identity — Method

delete_microsoft_teams_user_identity(chat_configuration_arn, user_id)
delete_microsoft_teams_user_identity(chat_configuration_arn, user_id, params::Dict{String,<:Any})

Deletes a Teams user identity

Arguments

chat_configuration_arn: The ARN of the MicrosoftTeamsChannelConfiguration associated with the user identity to delete.

user_id: Id from Microsoft Teams for user.

Main.Chatbot.delete_slack_channel_configuration — Method

delete_slack_channel_configuration(chat_configuration_arn)
delete_slack_channel_configuration(chat_configuration_arn, params::Dict{String,<:Any})

Deletes Slack Channel Configuration

Arguments

chat_configuration_arn: The ARN of the SlackChannelConfiguration to delete.

Main.Chatbot.delete_slack_user_identity — Method

delete_slack_user_identity(chat_configuration_arn, slack_team_id, slack_user_id)
delete_slack_user_identity(chat_configuration_arn, slack_team_id, slack_user_id, params::Dict{String,<:Any})

Deletes a Slack user identity

Arguments

chat_configuration_arn: The ARN of the SlackChannelConfiguration associated with the user identity to delete.

slack_team_id: The ID of the Slack workspace authorized with AWS Chatbot.

slack_user_id: The ID of the user in Slack.

Main.Chatbot.delete_slack_workspace_authorization — Method

delete_slack_workspace_authorization(slack_team_id)
delete_slack_workspace_authorization(slack_team_id, params::Dict{String,<:Any})

Deletes the Slack workspace authorization that allows channels to be configured in that workspace. This requires all configured channels in the workspace to be deleted.

Arguments

slack_team_id: The ID of the Slack workspace authorized with AWS Chatbot.

Main.Chatbot.describe_chime_webhook_configurations — Method

describe_chime_webhook_configurations()
describe_chime_webhook_configurations(params::Dict{String,<:Any})

Lists Chime Webhook Configurations optionally filtered by ChatConfigurationArn

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ChatConfigurationArn": An optional ARN of a ChimeWebhookConfiguration to describe.

"MaxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"NextToken": An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

Main.Chatbot.describe_slack_channel_configurations — Method

describe_slack_channel_configurations()
describe_slack_channel_configurations(params::Dict{String,<:Any})

Lists Slack Channel Configurations optionally filtered by ChatConfigurationArn

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ChatConfigurationArn": An optional ARN of a SlackChannelConfiguration to describe.

"MaxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"NextToken": An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

Main.Chatbot.describe_slack_user_identities — Method

describe_slack_user_identities()
describe_slack_user_identities(params::Dict{String,<:Any})

Lists all Slack user identities with a mapped role.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ChatConfigurationArn": The ARN of the SlackChannelConfiguration associated with the user identities to describe.

"MaxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"NextToken": An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

Main.Chatbot.describe_slack_workspaces — Method

describe_slack_workspaces()
describe_slack_workspaces(params::Dict{String,<:Any})

Lists all authorized Slack Workspaces for AWS Account

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"NextToken": An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

Main.Chatbot.get_account_preferences — Method

get_account_preferences()
get_account_preferences(params::Dict{String,<:Any})

Get Chatbot account level preferences

Main.Chatbot.get_microsoft_teams_channel_configuration — Method

get_microsoft_teams_channel_configuration(chat_configuration_arn)
get_microsoft_teams_channel_configuration(chat_configuration_arn, params::Dict{String,<:Any})

Get a single MS Teams Channel Configurations

Arguments

chat_configuration_arn: The ARN of the MicrosoftTeamsChannelConfiguration to retrieve.

Main.Chatbot.list_microsoft_teams_channel_configurations — Method

list_microsoft_teams_channel_configurations()
list_microsoft_teams_channel_configurations(params::Dict{String,<:Any})

Lists MS Teams Channel Configurations optionally filtered by TeamId

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"NextToken": An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

"TeamId": The ID of the Microsoft Team authorized with AWS Chatbot. To get the team ID, you must perform the initial authorization flow with Microsoft Teams in the AWS Chatbot console. Then you can copy and paste the team ID from the console. For more details, see steps 1-4 in Get started with Microsoft Teams in the AWS Chatbot Administrator Guide.

Main.Chatbot.list_microsoft_teams_configured_teams — Method

list_microsoft_teams_configured_teams()
list_microsoft_teams_configured_teams(params::Dict{String,<:Any})

Lists all authorized MS teams for AWS Account

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"NextToken": An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

Main.Chatbot.list_microsoft_teams_user_identities — Method

list_microsoft_teams_user_identities()
list_microsoft_teams_user_identities(params::Dict{String,<:Any})

Lists all Microsoft Teams user identities with a mapped role.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ChatConfigurationArn": The ARN of the MicrosoftTeamsChannelConfiguration associated with the user identities to list.

"MaxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"NextToken": An optional token returned from a prior request. Use this token for pagination of results from this action. If this parameter is specified, the response includes only results beyond the token, up to the value specified by MaxResults.

Main.Chatbot.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Retrieves the list of tags applied to a configuration.

Arguments

resource_arn: The ARN of the configuration.

Main.Chatbot.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Applies the supplied tags to a configuration.

Arguments

resource_arn: The ARN of the configuration.

tags: A list of tags to apply to the configuration.

Main.Chatbot.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes the supplied tags from a configuration

Arguments

resource_arn: The ARN of the configuration.

tag_keys: A list of tag keys to remove from the configuration.

Main.Chatbot.update_account_preferences — Method

update_account_preferences()
update_account_preferences(params::Dict{String,<:Any})

Update Chatbot account level preferences

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"TrainingDataCollectionEnabled": Turns on training data collection. This helps improve the AWS Chatbot experience by allowing AWS Chatbot to store and use your customer information, such as AWS Chatbot configurations, notifications, user inputs, AWS Chatbot generated responses, and interaction data. This data helps us to continuously improve and develop Artificial Intelligence (AI) technologies. Your data is not shared with any third parties and is protected using sophisticated controls to prevent unauthorized access and misuse. AWS Chatbot does not store or use interactions in chat channels with Amazon Q for training AWS Chatbot’s AI technologies.

"UserAuthorizationRequired": Enables use of a user role requirement in your chat configuration.

Main.Chatbot.update_chime_webhook_configuration — Method

update_chime_webhook_configuration(chat_configuration_arn)
update_chime_webhook_configuration(chat_configuration_arn, params::Dict{String,<:Any})

Updates a Chime Webhook Configuration

Arguments

chat_configuration_arn: The ARN of the ChimeWebhookConfiguration to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"IamRoleArn": The ARN of the IAM role that defines the permissions for AWS Chatbot. This is a user-defined role that AWS Chatbot will assume. This is not the service-linked role. For more information, see IAM Policies for AWS Chatbot.

"LoggingLevel": Logging levels include ERROR, INFO, or NONE.

"SnsTopicArns": The ARNs of the SNS topics that deliver notifications to AWS Chatbot.

"WebhookDescription": Description of the webhook. Recommend using the convention RoomName/WebhookName. See Chime setup tutorial for more details: https://docs.aws.amazon.com/chatbot/latest/adminguide/chime-setup.html.

"WebhookUrl": URL for the Chime webhook.

Main.Chatbot.update_microsoft_teams_channel_configuration — Method

update_microsoft_teams_channel_configuration(channel_id, chat_configuration_arn)
update_microsoft_teams_channel_configuration(channel_id, chat_configuration_arn, params::Dict{String,<:Any})

Updates MS Teams Channel Configuration

Arguments

channel_id: The ID of the Microsoft Teams channel.

chat_configuration_arn: The ARN of the MicrosoftTeamsChannelConfiguration to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ChannelName": The name of the Microsoft Teams channel.

"GuardrailPolicyArns": The list of IAM policy ARNs that are applied as channel guardrails. The AWS managed 'AdministratorAccess' policy is applied by default if this is not set.

"IamRoleArn": The ARN of the IAM role that defines the permissions for AWS Chatbot. This is a user-defined role that AWS Chatbot will assume. This is not the service-linked role. For more information, see IAM Policies for AWS Chatbot.

"LoggingLevel": Logging levels include ERROR, INFO, or NONE.

"SnsTopicArns": The ARNs of the SNS topics that deliver notifications to AWS Chatbot.

"UserAuthorizationRequired": Enables use of a user role requirement in your chat configuration.

Main.Chatbot.update_slack_channel_configuration — Method

update_slack_channel_configuration(chat_configuration_arn, slack_channel_id)
update_slack_channel_configuration(chat_configuration_arn, slack_channel_id, params::Dict{String,<:Any})

Updates Slack Channel Configuration

Arguments

chat_configuration_arn: The ARN of the SlackChannelConfiguration to update.

slack_channel_id: The ID of the Slack channel. To get the ID, open Slack, right click on the channel name in the left pane, then choose Copy Link. The channel ID is the 9-character string at the end of the URL. For example, ABCBBLZZZ.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"GuardrailPolicyArns": The list of IAM policy ARNs that are applied as channel guardrails. The AWS managed 'AdministratorAccess' policy is applied by default if this is not set.

"IamRoleArn": The ARN of the IAM role that defines the permissions for AWS Chatbot. This is a user-defined role that AWS Chatbot will assume. This is not the service-linked role. For more information, see IAM Policies for AWS Chatbot.

"LoggingLevel": Logging levels include ERROR, INFO, or NONE.

"SlackChannelName": The name of the Slack Channel.

"SnsTopicArns": The ARNs of the SNS topics that deliver notifications to AWS Chatbot.

"UserAuthorizationRequired": Enables use of a user role requirement in your chat configuration.