Connectparticipant

This page documents function available when using the Connectparticipant module, created with @service Connectparticipant.

Index

Main.Connectparticipant.complete_attachment_upload

Main.Connectparticipant.create_participant_connection

Main.Connectparticipant.disconnect_participant

Main.Connectparticipant.get_attachment

Main.Connectparticipant.get_transcript

Main.Connectparticipant.send_event

Main.Connectparticipant.send_message

Main.Connectparticipant.start_attachment_upload

Documentation

Main.Connectparticipant.complete_attachment_upload — Method

complete_attachment_upload(attachment_ids, client_token, x-_amz-_bearer)
complete_attachment_upload(attachment_ids, client_token, x-_amz-_bearer, params::Dict{String,<:Any})

Allows you to confirm that the attachment has been uploaded using the pre-signed URL provided in StartAttachmentUpload API.

Arguments

attachment_ids: A list of unique identifiers for the attachments.

client_token: A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

x-_amz-_bearer: The authentication token associated with the participant's connection.

Main.Connectparticipant.create_participant_connection — Method

create_participant_connection(type, x-_amz-_bearer)
create_participant_connection(type, x-_amz-_bearer, params::Dict{String,<:Any})

Creates the participant's connection. Note that ParticipantToken is used for invoking this API instead of ConnectionToken. The participant token is valid for the lifetime of the participant – until they are part of a contact. The response URL for WEBSOCKET Type has a connect expiry timeout of 100s. Clients must manually connect to the returned websocket URL and subscribe to the desired topic. For chat, you need to publish the following on the established websocket connection: {"topic":"aws/subscribe","content":{"topics":["aws/chat"]}} Upon websocket URL expiry, as specified in the response ConnectionExpiry parameter, clients need to call this API again to obtain a new websocket URL and perform the same steps as before. The Amazon Connect Participant Service APIs do not use Signature Version 4 authentication.

Arguments

type: Type of connection information required.

x-_amz-_bearer: This is a header parameter. The Participant Token as obtained from StartChatContact API response.

Main.Connectparticipant.disconnect_participant — Method

disconnect_participant(x-_amz-_bearer)
disconnect_participant(x-_amz-_bearer, params::Dict{String,<:Any})

Disconnects a participant. Note that ConnectionToken is used for invoking this API instead of ParticipantToken. The Amazon Connect Participant Service APIs do not use Signature Version 4 authentication.

Arguments

x-_amz-_bearer: The authentication token associated with the participant's connection.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

Main.Connectparticipant.get_attachment — Method

get_attachment(attachment_id, x-_amz-_bearer)
get_attachment(attachment_id, x-_amz-_bearer, params::Dict{String,<:Any})

Provides a pre-signed URL for download of a completed attachment. This is an asynchronous API for use with active contacts.

Arguments

attachment_id: A unique identifier for the attachment.

x-_amz-_bearer: The authentication token associated with the participant's connection.

Main.Connectparticipant.get_transcript — Method

get_transcript(x-_amz-_bearer)
get_transcript(x-_amz-_bearer, params::Dict{String,<:Any})

Retrieves a transcript of the session, including details about any attachments. Note that ConnectionToken is used for invoking this API instead of ParticipantToken. The Amazon Connect Participant Service APIs do not use Signature Version 4 authentication.

Arguments

x-_amz-_bearer: The authentication token associated with the participant's connection.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ContactId": The contactId from the current contact chain for which transcript is needed.

"MaxResults": The maximum number of results to return in the page. Default: 10.

"NextToken": The pagination token. Use the value returned previously in the next subsequent request to retrieve the next set of results.

"ScanDirection": The direction from StartPosition from which to retrieve message. Default: BACKWARD when no StartPosition is provided, FORWARD with StartPosition.

"SortOrder": The sort order for the records. Default: DESCENDING.

"StartPosition": A filtering option for where to start.

Main.Connectparticipant.send_event — Method

send_event(content_type, x-_amz-_bearer)
send_event(content_type, x-_amz-_bearer, params::Dict{String,<:Any})

Sends an event. Note that ConnectionToken is used for invoking this API instead of ParticipantToken. The Amazon Connect Participant Service APIs do not use Signature Version 4 authentication.

Arguments

content_type: The content type of the request. Supported types are: application/vnd.amazonaws.connect.event.typing application/vnd.amazonaws.connect.event.connection.acknowledged

x-_amz-_bearer: The authentication token associated with the participant's connection.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

"Content": The content of the event to be sent (for example, message text). This is not yet supported.

Main.Connectparticipant.send_message — Method

send_message(content, content_type, x-_amz-_bearer)
send_message(content, content_type, x-_amz-_bearer, params::Dict{String,<:Any})

Sends a message. Note that ConnectionToken is used for invoking this API instead of ParticipantToken. The Amazon Connect Participant Service APIs do not use Signature Version 4 authentication.

Arguments

content: The content of the message.

content_type: The type of the content. Supported types are text/plain.

x-_amz-_bearer: The authentication token associated with the connection.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

Main.Connectparticipant.start_attachment_upload — Method

start_attachment_upload(attachment_name, attachment_size_in_bytes, client_token, content_type, x-_amz-_bearer)
start_attachment_upload(attachment_name, attachment_size_in_bytes, client_token, content_type, x-_amz-_bearer, params::Dict{String,<:Any})

Provides a pre-signed Amazon S3 URL in response for uploading the file directly to S3.

Arguments

attachment_name: A case-sensitive name of the attachment being uploaded.

attachment_size_in_bytes: The size of the attachment in bytes.

client_token: A unique case sensitive identifier to support idempotency of request.

content_type: Describes the MIME file type of the attachment. For a list of supported file types, see Feature specifications in the Amazon Connect Administrator Guide.

x-_amz-_bearer: The authentication token associated with the participant's connection.