Elasticsearch Service

This page documents function available when using the Elasticsearch_Service module, created with @service Elasticsearch_Service.

Index

Main.Elasticsearch_Service.accept_inbound_cross_cluster_search_connection

Main.Elasticsearch_Service.add_tags

Main.Elasticsearch_Service.associate_package

Main.Elasticsearch_Service.cancel_elasticsearch_service_software_update

Main.Elasticsearch_Service.create_elasticsearch_domain

Main.Elasticsearch_Service.create_outbound_cross_cluster_search_connection

Main.Elasticsearch_Service.create_package

Main.Elasticsearch_Service.delete_elasticsearch_domain

Main.Elasticsearch_Service.delete_elasticsearch_service_role

Main.Elasticsearch_Service.delete_inbound_cross_cluster_search_connection

Main.Elasticsearch_Service.delete_outbound_cross_cluster_search_connection

Main.Elasticsearch_Service.delete_package

Main.Elasticsearch_Service.describe_elasticsearch_domain

Main.Elasticsearch_Service.describe_elasticsearch_domain_config

Main.Elasticsearch_Service.describe_elasticsearch_domains

Main.Elasticsearch_Service.describe_elasticsearch_instance_type_limits

Main.Elasticsearch_Service.describe_inbound_cross_cluster_search_connections

Main.Elasticsearch_Service.describe_outbound_cross_cluster_search_connections

Main.Elasticsearch_Service.describe_packages

Main.Elasticsearch_Service.describe_reserved_elasticsearch_instance_offerings

Main.Elasticsearch_Service.describe_reserved_elasticsearch_instances

Main.Elasticsearch_Service.dissociate_package

Main.Elasticsearch_Service.get_compatible_elasticsearch_versions

Main.Elasticsearch_Service.get_package_version_history

Main.Elasticsearch_Service.get_upgrade_history

Main.Elasticsearch_Service.get_upgrade_status

Main.Elasticsearch_Service.list_domain_names

Main.Elasticsearch_Service.list_domains_for_package

Main.Elasticsearch_Service.list_elasticsearch_instance_types

Main.Elasticsearch_Service.list_elasticsearch_versions

Main.Elasticsearch_Service.list_packages_for_domain

Main.Elasticsearch_Service.list_tags

Main.Elasticsearch_Service.purchase_reserved_elasticsearch_instance_offering

Main.Elasticsearch_Service.reject_inbound_cross_cluster_search_connection

Main.Elasticsearch_Service.remove_tags

Main.Elasticsearch_Service.start_elasticsearch_service_software_update

Main.Elasticsearch_Service.update_elasticsearch_domain_config

Main.Elasticsearch_Service.update_package

Main.Elasticsearch_Service.upgrade_elasticsearch_domain

Documentation

Main.Elasticsearch_Service.accept_inbound_cross_cluster_search_connection — Method

AcceptInboundCrossClusterSearchConnection()

Allows the destination domain owner to accept an inbound cross-cluster search connection request.

Required Parameters

ConnectionId: The id of the inbound connection that you want to accept.

Main.Elasticsearch_Service.add_tags — Method

AddTags()

Attaches tags to an existing Elasticsearch domain. Tags are a set of case-sensitive key value pairs. An Elasticsearch domain may have up to 10 tags. See Tagging Amazon Elasticsearch Service Domains for more information.

Required Parameters

ARN: Specify the ARN for which you want to add the tags.

TagList: List of Tag that need to be added for the Elasticsearch domain.

Main.Elasticsearch_Service.associate_package — Method

AssociatePackage()

Associates a package with an Amazon ES domain.

Required Parameters

DomainName: Name of the domain that you want to associate the package with.

PackageID: Internal ID of the package that you want to associate with a domain. Use DescribePackages to find this value.

Main.Elasticsearch_Service.cancel_elasticsearch_service_software_update — Method

CancelElasticsearchServiceSoftwareUpdate()

Cancels a scheduled service software update for an Amazon ES domain. You can only perform this operation before the AutomatedUpdateDate and when the UpdateStatus is in the PENDING_UPDATE state.

Required Parameters

DomainName: The name of the domain that you want to stop the latest service software update on.

Main.Elasticsearch_Service.create_elasticsearch_domain — Method

CreateElasticsearchDomain()

Creates a new Elasticsearch domain. For more information, see Creating Elasticsearch Domains in the Amazon Elasticsearch Service Developer Guide.

Required Parameters

DomainName: The name of the Elasticsearch domain that you are creating. Domain names are unique across the domains owned by an account within an AWS region. Domain names must start with a lowercase letter and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

Optional Parameters

AccessPolicies: IAM access policy as a JSON-formatted string.

AdvancedOptions: Option to allow references to indices in an HTTP request body. Must be false when configuring access to individual sub-resources. By default, the value is true. See Configuration Advanced Options for more information.

AdvancedSecurityOptions: Specifies advanced security options.

CognitoOptions: Options to specify the Cognito user and identity pools for Kibana authentication. For more information, see Amazon Cognito Authentication for Kibana.

DomainEndpointOptions: Options to specify configuration that will be applied to the domain endpoint.

EBSOptions: Options to enable, disable and specify the type and size of EBS storage volumes.

ElasticsearchClusterConfig: Configuration options for an Elasticsearch domain. Specifies the instance type and number of instances in the domain cluster.

ElasticsearchVersion: String of format X.Y to specify version for the Elasticsearch domain eg. "1.5" or "2.3". For more information, see Creating Elasticsearch Domains in the Amazon Elasticsearch Service Developer Guide.

EncryptionAtRestOptions: Specifies the Encryption At Rest Options.

LogPublishingOptions: Map of LogType and LogPublishingOption, each containing options to publish a given type of Elasticsearch log.

NodeToNodeEncryptionOptions: Specifies the NodeToNodeEncryptionOptions.

SnapshotOptions: Option to set time, in UTC format, of the daily automated snapshot. Default value is 0 hours.

VPCOptions: Options to specify the subnets and security groups for VPC endpoint. For more information, see Creating a VPC in VPC Endpoints for Amazon Elasticsearch Service Domains

Main.Elasticsearch_Service.create_outbound_cross_cluster_search_connection — Method

CreateOutboundCrossClusterSearchConnection()

Creates a new cross-cluster search connection from a source domain to a destination domain.

Required Parameters

ConnectionAlias: Specifies the connection alias that will be used by the customer for this connection.

DestinationDomainInfo: Specifies the DomainInformation for the destination Elasticsearch domain.

SourceDomainInfo: Specifies the DomainInformation for the source Elasticsearch domain.

Main.Elasticsearch_Service.create_package — Method

CreatePackage()

Create a package for use with Amazon ES domains.

Required Parameters

PackageName: Unique identifier for the package.

PackageSource: The customer S3 location PackageSource for importing the package.

PackageType: Type of package. Currently supports only TXT-DICTIONARY.

Optional Parameters

PackageDescription: Description of the package.

Main.Elasticsearch_Service.delete_elasticsearch_domain — Method

DeleteElasticsearchDomain()

Permanently deletes the specified Elasticsearch domain and all of its data. Once a domain is deleted, it cannot be recovered.

Required Parameters

DomainName: The name of the Elasticsearch domain that you want to permanently delete.

Main.Elasticsearch_Service.delete_elasticsearch_service_role — Method

DeleteElasticsearchServiceRole()

Deletes the service-linked role that Elasticsearch Service uses to manage and maintain VPC domains. Role deletion will fail if any existing VPC domains use the role. You must delete any such Elasticsearch domains before deleting the role. See Deleting Elasticsearch Service Role in VPC Endpoints for Amazon Elasticsearch Service Domains.

Main.Elasticsearch_Service.delete_inbound_cross_cluster_search_connection — Method

DeleteInboundCrossClusterSearchConnection()

Allows the destination domain owner to delete an existing inbound cross-cluster search connection.

Required Parameters

ConnectionId: The id of the inbound connection that you want to permanently delete.

Main.Elasticsearch_Service.delete_outbound_cross_cluster_search_connection — Method

DeleteOutboundCrossClusterSearchConnection()

Allows the source domain owner to delete an existing outbound cross-cluster search connection.

Required Parameters

ConnectionId: The id of the outbound connection that you want to permanently delete.

Main.Elasticsearch_Service.delete_package — Method

DeletePackage()

Delete the package.

Required Parameters

PackageID: Internal ID of the package that you want to delete. Use DescribePackages to find this value.

Main.Elasticsearch_Service.describe_elasticsearch_domain — Method

DescribeElasticsearchDomain()

Returns domain configuration information about the specified Elasticsearch domain, including the domain ID, domain endpoint, and domain ARN.

Required Parameters

DomainName: The name of the Elasticsearch domain for which you want information.

Main.Elasticsearch_Service.describe_elasticsearch_domain_config — Method

DescribeElasticsearchDomainConfig()

Provides cluster configuration information about the specified Elasticsearch domain, such as the state, creation date, update version, and update date for cluster options.

Required Parameters

DomainName: The Elasticsearch domain that you want to get information about.

Main.Elasticsearch_Service.describe_elasticsearch_domains — Method

DescribeElasticsearchDomains()

Returns domain configuration information about the specified Elasticsearch domains, including the domain ID, domain endpoint, and domain ARN.

Required Parameters

DomainNames: The Elasticsearch domains for which you want information.

Main.Elasticsearch_Service.describe_elasticsearch_instance_type_limits — Method

DescribeElasticsearchInstanceTypeLimits()

Describe Elasticsearch Limits for a given InstanceType and ElasticsearchVersion. When modifying existing Domain, specify the DomainName to know what Limits are supported for modifying.

Required Parameters

ElasticsearchVersion: Version of Elasticsearch for which Limits are needed.

InstanceType: The instance type for an Elasticsearch cluster for which Elasticsearch Limits are needed.

Optional Parameters

domainName: DomainName represents the name of the Domain that we are trying to modify. This should be present only if we are querying for Elasticsearch Limits for existing domain.

Main.Elasticsearch_Service.describe_inbound_cross_cluster_search_connections — Method

DescribeInboundCrossClusterSearchConnections()

Lists all the inbound cross-cluster search connections for a destination domain.

Optional Parameters

Filters: A list of filters used to match properties for inbound cross-cluster search connection. Available Filter names for this operation are: cross-cluster-search-connection-id source-domain-info.domain-name source-domain-info.owner-id source-domain-info.region destination-domain-info.domain-name

MaxResults: Set this value to limit the number of results returned. If not specified, defaults to 100.

NextToken: NextToken is sent in case the earlier API call results contain the NextToken. It is used for pagination.

Main.Elasticsearch_Service.describe_outbound_cross_cluster_search_connections — Method

DescribeOutboundCrossClusterSearchConnections()

Lists all the outbound cross-cluster search connections for a source domain.

Optional Parameters

Filters: A list of filters used to match properties for outbound cross-cluster search connection. Available Filter names for this operation are: cross-cluster-search-connection-id destination-domain-info.domain-name destination-domain-info.owner-id destination-domain-info.region source-domain-info.domain-name

MaxResults: Set this value to limit the number of results returned. If not specified, defaults to 100.

NextToken: NextToken is sent in case the earlier API call results contain the NextToken. It is used for pagination.

Main.Elasticsearch_Service.describe_packages — Method

DescribePackages()

Describes all packages available to Amazon ES. Includes options for filtering, limiting the number of results, and pagination.

Optional Parameters

Filters: Only returns packages that match the DescribePackagesFilterList values.

MaxResults: Limits results to a maximum number of packages.

NextToken: Used for pagination. Only necessary if a previous API call includes a non-null NextToken value. If provided, returns results for the next page.

Main.Elasticsearch_Service.describe_reserved_elasticsearch_instance_offerings — Method

DescribeReservedElasticsearchInstanceOfferings()

Lists available reserved Elasticsearch instance offerings.

Optional Parameters

maxResults: Set this value to limit the number of results returned. If not specified, defaults to 100.

nextToken: NextToken should be sent in case if earlier API call produced result containing NextToken. It is used for pagination.

offeringId: The offering identifier filter value. Use this parameter to show only the available offering that matches the specified reservation identifier.

Main.Elasticsearch_Service.describe_reserved_elasticsearch_instances — Method

DescribeReservedElasticsearchInstances()

Returns information about reserved Elasticsearch instances for this account.

Optional Parameters

maxResults: Set this value to limit the number of results returned. If not specified, defaults to 100.

nextToken: NextToken should be sent in case if earlier API call produced result containing NextToken. It is used for pagination.

reservationId: The reserved instance identifier filter value. Use this parameter to show only the reservation that matches the specified reserved Elasticsearch instance ID.

Main.Elasticsearch_Service.dissociate_package — Method

DissociatePackage()

Dissociates a package from the Amazon ES domain.

Required Parameters

DomainName: Name of the domain that you want to associate the package with.

PackageID: Internal ID of the package that you want to associate with a domain. Use DescribePackages to find this value.

Main.Elasticsearch_Service.get_compatible_elasticsearch_versions — Method

GetCompatibleElasticsearchVersions()

Returns a list of upgrade compatible Elastisearch versions. You can optionally pass a DomainName to get all upgrade compatible Elasticsearch versions for that specific domain.

Optional Parameters

domainName:

Main.Elasticsearch_Service.get_package_version_history — Method

GetPackageVersionHistory()

Returns a list of versions of the package, along with their creation time and commit message.

Required Parameters

PackageID: Returns an audit history of versions of the package.

Optional Parameters

maxResults: Limits results to a maximum number of versions.

nextToken: Used for pagination. Only necessary if a previous API call includes a non-null NextToken value. If provided, returns results for the next page.

Main.Elasticsearch_Service.get_upgrade_history — Method

GetUpgradeHistory()

Retrieves the complete history of the last 10 upgrades that were performed on the domain.

Required Parameters

DomainName:

Optional Parameters

maxResults:

nextToken:

Main.Elasticsearch_Service.get_upgrade_status — Method

GetUpgradeStatus()

Retrieves the latest status of the last upgrade or upgrade eligibility check that was performed on the domain.

Required Parameters

DomainName:

Main.Elasticsearch_Service.list_domain_names — Method

ListDomainNames()

Returns the name of all Elasticsearch domains owned by the current user's account.

Main.Elasticsearch_Service.list_domains_for_package — Method

ListDomainsForPackage()

Lists all Amazon ES domains associated with the package.

Required Parameters

PackageID: The package for which to list domains.

Optional Parameters

maxResults: Limits results to a maximum number of domains.

nextToken: Used for pagination. Only necessary if a previous API call includes a non-null NextToken value. If provided, returns results for the next page.

Main.Elasticsearch_Service.list_elasticsearch_instance_types — Method

ListElasticsearchInstanceTypes()

List all Elasticsearch instance types that are supported for given ElasticsearchVersion

Required Parameters

ElasticsearchVersion: Version of Elasticsearch for which list of supported elasticsearch instance types are needed.

Optional Parameters

domainName: DomainName represents the name of the Domain that we are trying to modify. This should be present only if we are querying for list of available Elasticsearch instance types when modifying existing domain.

maxResults: Set this value to limit the number of results returned. Value provided must be greater than 30 else it wont be honored.

nextToken: NextToken should be sent in case if earlier API call produced result containing NextToken. It is used for pagination.

Main.Elasticsearch_Service.list_elasticsearch_versions — Method

ListElasticsearchVersions()

List all supported Elasticsearch versions

Optional Parameters

maxResults: Set this value to limit the number of results returned. Value provided must be greater than 10 else it wont be honored.

nextToken:

Main.Elasticsearch_Service.list_packages_for_domain — Method

ListPackagesForDomain()

Lists all packages associated with the Amazon ES domain.

Required Parameters

DomainName: The name of the domain for which you want to list associated packages.

Optional Parameters

maxResults: Limits results to a maximum number of packages.

nextToken: Used for pagination. Only necessary if a previous API call includes a non-null NextToken value. If provided, returns results for the next page.

Main.Elasticsearch_Service.list_tags — Method

ListTags()

Returns all tags for the given Elasticsearch domain.

Required Parameters

arn: Specify the ARN for the Elasticsearch domain to which the tags are attached that you want to view.

Main.Elasticsearch_Service.purchase_reserved_elasticsearch_instance_offering — Method

PurchaseReservedElasticsearchInstanceOffering()

Allows you to purchase reserved Elasticsearch instances.

Required Parameters

ReservationName: A customer-specified identifier to track this reservation.

ReservedElasticsearchInstanceOfferingId: The ID of the reserved Elasticsearch instance offering to purchase.

Optional Parameters

InstanceCount: The number of Elasticsearch instances to reserve.

Main.Elasticsearch_Service.reject_inbound_cross_cluster_search_connection — Method

RejectInboundCrossClusterSearchConnection()

Allows the destination domain owner to reject an inbound cross-cluster search connection request.

Required Parameters

ConnectionId: The id of the inbound connection that you want to reject.

Main.Elasticsearch_Service.remove_tags — Method

RemoveTags()

Removes the specified set of tags from the specified Elasticsearch domain.

Required Parameters

ARN: Specifies the ARN for the Elasticsearch domain from which you want to delete the specified tags.

TagKeys: Specifies the TagKey list which you want to remove from the Elasticsearch domain.

Main.Elasticsearch_Service.start_elasticsearch_service_software_update — Method

StartElasticsearchServiceSoftwareUpdate()

Schedules a service software update for an Amazon ES domain.

Required Parameters

DomainName: The name of the domain that you want to update to the latest service software.

Main.Elasticsearch_Service.update_elasticsearch_domain_config — Method

UpdateElasticsearchDomainConfig()

Modifies the cluster configuration of the specified Elasticsearch domain, setting as setting the instance type and the number of instances.

Required Parameters

DomainName: The name of the Elasticsearch domain that you are updating.

Optional Parameters

AccessPolicies: IAM access policy as a JSON-formatted string.

AdvancedOptions: Modifies the advanced option to allow references to indices in an HTTP request body. Must be false when configuring access to individual sub-resources. By default, the value is true. See Configuration Advanced Options for more information.

AdvancedSecurityOptions: Specifies advanced security options.

CognitoOptions: Options to specify the Cognito user and identity pools for Kibana authentication. For more information, see Amazon Cognito Authentication for Kibana.

DomainEndpointOptions: Options to specify configuration that will be applied to the domain endpoint.

EBSOptions: Specify the type and size of the EBS volume that you want to use.

ElasticsearchClusterConfig: The type and number of instances to instantiate for the domain cluster.

EncryptionAtRestOptions: Specifies the Encryption At Rest Options.

LogPublishingOptions: Map of LogType and LogPublishingOption, each containing options to publish a given type of Elasticsearch log.

NodeToNodeEncryptionOptions: Specifies the NodeToNodeEncryptionOptions.

SnapshotOptions: Option to set the time, in UTC format, for the daily automated snapshot. Default value is 0 hours.

VPCOptions: Options to specify the subnets and security groups for VPC endpoint. For more information, see Creating a VPC in VPC Endpoints for Amazon Elasticsearch Service Domains

Main.Elasticsearch_Service.update_package — Method

UpdatePackage()

Updates a package for use with Amazon ES domains.

Required Parameters

PackageID: Unique identifier for the package.

PackageSource:

Optional Parameters

CommitMessage: An info message for the new version which will be shown as part of GetPackageVersionHistoryResponse.

PackageDescription: New description of the package.

Main.Elasticsearch_Service.upgrade_elasticsearch_domain — Method

UpgradeElasticsearchDomain()

Allows you to either upgrade your domain or perform an Upgrade eligibility check to a compatible Elasticsearch version.

Required Parameters

DomainName:

TargetVersion: The version of Elasticsearch that you intend to upgrade the domain to.

Optional Parameters

PerformCheckOnly: This flag, when set to True, indicates that an Upgrade Eligibility Check needs to be performed. This will not actually perform the Upgrade.