Elasticsearch Service

This page documents function available when using the Elasticsearch_Service module, created with @service Elasticsearch_Service.

Index

Main.Elasticsearch_Service.accept_inbound_cross_cluster_search_connection

Main.Elasticsearch_Service.add_tags

Main.Elasticsearch_Service.associate_package

Main.Elasticsearch_Service.authorize_vpc_endpoint_access

Main.Elasticsearch_Service.cancel_elasticsearch_service_software_update

Main.Elasticsearch_Service.create_elasticsearch_domain

Main.Elasticsearch_Service.create_outbound_cross_cluster_search_connection

Main.Elasticsearch_Service.create_package

Main.Elasticsearch_Service.create_vpc_endpoint

Main.Elasticsearch_Service.delete_elasticsearch_domain

Main.Elasticsearch_Service.delete_elasticsearch_service_role

Main.Elasticsearch_Service.delete_inbound_cross_cluster_search_connection

Main.Elasticsearch_Service.delete_outbound_cross_cluster_search_connection

Main.Elasticsearch_Service.delete_package

Main.Elasticsearch_Service.delete_vpc_endpoint

Main.Elasticsearch_Service.describe_domain_auto_tunes

Main.Elasticsearch_Service.describe_domain_change_progress

Main.Elasticsearch_Service.describe_elasticsearch_domain

Main.Elasticsearch_Service.describe_elasticsearch_domain_config

Main.Elasticsearch_Service.describe_elasticsearch_domains

Main.Elasticsearch_Service.describe_elasticsearch_instance_type_limits

Main.Elasticsearch_Service.describe_inbound_cross_cluster_search_connections

Main.Elasticsearch_Service.describe_outbound_cross_cluster_search_connections

Main.Elasticsearch_Service.describe_packages

Main.Elasticsearch_Service.describe_reserved_elasticsearch_instance_offerings

Main.Elasticsearch_Service.describe_reserved_elasticsearch_instances

Main.Elasticsearch_Service.describe_vpc_endpoints

Main.Elasticsearch_Service.dissociate_package

Main.Elasticsearch_Service.get_compatible_elasticsearch_versions

Main.Elasticsearch_Service.get_package_version_history

Main.Elasticsearch_Service.get_upgrade_history

Main.Elasticsearch_Service.get_upgrade_status

Main.Elasticsearch_Service.list_domain_names

Main.Elasticsearch_Service.list_domains_for_package

Main.Elasticsearch_Service.list_elasticsearch_instance_types

Main.Elasticsearch_Service.list_elasticsearch_versions

Main.Elasticsearch_Service.list_packages_for_domain

Main.Elasticsearch_Service.list_tags

Main.Elasticsearch_Service.list_vpc_endpoint_access

Main.Elasticsearch_Service.list_vpc_endpoints

Main.Elasticsearch_Service.list_vpc_endpoints_for_domain

Main.Elasticsearch_Service.purchase_reserved_elasticsearch_instance_offering

Main.Elasticsearch_Service.reject_inbound_cross_cluster_search_connection

Main.Elasticsearch_Service.remove_tags

Main.Elasticsearch_Service.revoke_vpc_endpoint_access

Main.Elasticsearch_Service.start_elasticsearch_service_software_update

Main.Elasticsearch_Service.update_elasticsearch_domain_config

Main.Elasticsearch_Service.update_package

Main.Elasticsearch_Service.update_vpc_endpoint

Main.Elasticsearch_Service.upgrade_elasticsearch_domain

Documentation

Main.Elasticsearch_Service.accept_inbound_cross_cluster_search_connection — Method

accept_inbound_cross_cluster_search_connection(connection_id)
accept_inbound_cross_cluster_search_connection(connection_id, params::Dict{String,<:Any})

Allows the destination domain owner to accept an inbound cross-cluster search connection request.

Arguments

connection_id: The id of the inbound connection that you want to accept.

Main.Elasticsearch_Service.add_tags — Method

add_tags(arn, tag_list)
add_tags(arn, tag_list, params::Dict{String,<:Any})

Attaches tags to an existing Elasticsearch domain. Tags are a set of case-sensitive key value pairs. An Elasticsearch domain may have up to 10 tags. See Tagging Amazon Elasticsearch Service Domains for more information.

Arguments

arn: Specify the ARN for which you want to add the tags.

tag_list: List of Tag that need to be added for the Elasticsearch domain.

Main.Elasticsearch_Service.associate_package — Method

associate_package(domain_name, package_id)
associate_package(domain_name, package_id, params::Dict{String,<:Any})

Associates a package with an Amazon ES domain.

Arguments

domain_name: Name of the domain that you want to associate the package with.

package_id: Internal ID of the package that you want to associate with a domain. Use DescribePackages to find this value.

Main.Elasticsearch_Service.authorize_vpc_endpoint_access — Method

authorize_vpc_endpoint_access(account, domain_name)
authorize_vpc_endpoint_access(account, domain_name, params::Dict{String,<:Any})

Provides access to an Amazon OpenSearch Service domain through the use of an interface VPC endpoint.

Arguments

account: The account ID to grant access to.

domain_name: The name of the OpenSearch Service domain to provide access to.

Main.Elasticsearch_Service.cancel_elasticsearch_service_software_update — Method

cancel_elasticsearch_service_software_update(domain_name)
cancel_elasticsearch_service_software_update(domain_name, params::Dict{String,<:Any})

Cancels a scheduled service software update for an Amazon ES domain. You can only perform this operation before the AutomatedUpdateDate and when the UpdateStatus is in the PENDING_UPDATE state.

Arguments

domain_name: The name of the domain that you want to stop the latest service software update on.

Main.Elasticsearch_Service.create_elasticsearch_domain — Method

create_elasticsearch_domain(domain_name)
create_elasticsearch_domain(domain_name, params::Dict{String,<:Any})

Creates a new Elasticsearch domain. For more information, see Creating Elasticsearch Domains in the Amazon Elasticsearch Service Developer Guide.

Arguments

domain_name: The name of the Elasticsearch domain that you are creating. Domain names are unique across the domains owned by an account within an AWS region. Domain names must start with a lowercase letter and can contain the following characters: a-z (lowercase), 0-9, and - (hyphen).

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AccessPolicies": IAM access policy as a JSON-formatted string.

"AdvancedOptions": Option to allow references to indices in an HTTP request body. Must be false when configuring access to individual sub-resources. By default, the value is true. See Configuration Advanced Options for more information.

"AdvancedSecurityOptions": Specifies advanced security options.

"AutoTuneOptions": Specifies Auto-Tune options.

"CognitoOptions": Options to specify the Cognito user and identity pools for Kibana authentication. For more information, see Amazon Cognito Authentication for Kibana.

"DomainEndpointOptions": Options to specify configuration that will be applied to the domain endpoint.

"EBSOptions": Options to enable, disable and specify the type and size of EBS storage volumes.

"ElasticsearchClusterConfig": Configuration options for an Elasticsearch domain. Specifies the instance type and number of instances in the domain cluster.

"ElasticsearchVersion": String of format X.Y to specify version for the Elasticsearch domain eg. "1.5" or "2.3". For more information, see Creating Elasticsearch Domains in the Amazon Elasticsearch Service Developer Guide.

"EncryptionAtRestOptions": Specifies the Encryption At Rest Options.

"LogPublishingOptions": Map of LogType and LogPublishingOption, each containing options to publish a given type of Elasticsearch log.

"NodeToNodeEncryptionOptions": Specifies the NodeToNodeEncryptionOptions.

"SnapshotOptions": Option to set time, in UTC format, of the daily automated snapshot. Default value is 0 hours.

"TagList": A list of Tag added during domain creation.

"VPCOptions": Options to specify the subnets and security groups for VPC endpoint. For more information, see Creating a VPC in VPC Endpoints for Amazon Elasticsearch Service Domains

Main.Elasticsearch_Service.create_outbound_cross_cluster_search_connection — Method

create_outbound_cross_cluster_search_connection(connection_alias, destination_domain_info, source_domain_info)
create_outbound_cross_cluster_search_connection(connection_alias, destination_domain_info, source_domain_info, params::Dict{String,<:Any})

Creates a new cross-cluster search connection from a source domain to a destination domain.

Arguments

connection_alias: Specifies the connection alias that will be used by the customer for this connection.

destination_domain_info: Specifies the DomainInformation for the destination Elasticsearch domain.

source_domain_info: Specifies the DomainInformation for the source Elasticsearch domain.

Main.Elasticsearch_Service.create_package — Method

create_package(package_name, package_source, package_type)
create_package(package_name, package_source, package_type, params::Dict{String,<:Any})

Create a package for use with Amazon ES domains.

Arguments

package_name: Unique identifier for the package.

package_source: The customer S3 location PackageSource for importing the package.

package_type: Type of package. Currently supports only TXT-DICTIONARY.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"PackageDescription": Description of the package.

Main.Elasticsearch_Service.create_vpc_endpoint — Method

create_vpc_endpoint(domain_arn, vpc_options)
create_vpc_endpoint(domain_arn, vpc_options, params::Dict{String,<:Any})

Creates an Amazon OpenSearch Service-managed VPC endpoint.

Arguments

domain_arn: The Amazon Resource Name (ARN) of the domain to grant access to.

vpc_options: Options to specify the subnets and security groups for the endpoint.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientToken": Unique, case-sensitive identifier to ensure idempotency of the request.

Main.Elasticsearch_Service.delete_elasticsearch_domain — Method

delete_elasticsearch_domain(domain_name)
delete_elasticsearch_domain(domain_name, params::Dict{String,<:Any})

Permanently deletes the specified Elasticsearch domain and all of its data. Once a domain is deleted, it cannot be recovered.

Arguments

domain_name: The name of the Elasticsearch domain that you want to permanently delete.

Main.Elasticsearch_Service.delete_elasticsearch_service_role — Method

delete_elasticsearch_service_role()
delete_elasticsearch_service_role(params::Dict{String,<:Any})

Deletes the service-linked role that Elasticsearch Service uses to manage and maintain VPC domains. Role deletion will fail if any existing VPC domains use the role. You must delete any such Elasticsearch domains before deleting the role. See Deleting Elasticsearch Service Role in VPC Endpoints for Amazon Elasticsearch Service Domains.

Main.Elasticsearch_Service.delete_inbound_cross_cluster_search_connection — Method

delete_inbound_cross_cluster_search_connection(connection_id)
delete_inbound_cross_cluster_search_connection(connection_id, params::Dict{String,<:Any})

Allows the destination domain owner to delete an existing inbound cross-cluster search connection.

Arguments

connection_id: The id of the inbound connection that you want to permanently delete.

Main.Elasticsearch_Service.delete_outbound_cross_cluster_search_connection — Method

delete_outbound_cross_cluster_search_connection(connection_id)
delete_outbound_cross_cluster_search_connection(connection_id, params::Dict{String,<:Any})

Allows the source domain owner to delete an existing outbound cross-cluster search connection.

Arguments

connection_id: The id of the outbound connection that you want to permanently delete.

Main.Elasticsearch_Service.delete_package — Method

delete_package(package_id)
delete_package(package_id, params::Dict{String,<:Any})

Delete the package.

Arguments

package_id: Internal ID of the package that you want to delete. Use DescribePackages to find this value.

Main.Elasticsearch_Service.delete_vpc_endpoint — Method

delete_vpc_endpoint(vpc_endpoint_id)
delete_vpc_endpoint(vpc_endpoint_id, params::Dict{String,<:Any})

Deletes an Amazon OpenSearch Service-managed interface VPC endpoint.

Arguments

vpc_endpoint_id: The unique identifier of the endpoint to be deleted.

Main.Elasticsearch_Service.describe_domain_auto_tunes — Method

describe_domain_auto_tunes(domain_name)
describe_domain_auto_tunes(domain_name, params::Dict{String,<:Any})

Provides scheduled Auto-Tune action details for the Elasticsearch domain, such as Auto-Tune action type, description, severity, and scheduled date.

Arguments

domain_name: Specifies the domain name for which you want Auto-Tune action details.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": Set this value to limit the number of results returned. If not specified, defaults to 100.

"NextToken": NextToken is sent in case the earlier API call results contain the NextToken. It is used for pagination.

Main.Elasticsearch_Service.describe_domain_change_progress — Method

describe_domain_change_progress(domain_name)
describe_domain_change_progress(domain_name, params::Dict{String,<:Any})

Returns information about the current blue/green deployment happening on a domain, including a change ID, status, and progress stages.

Arguments

domain_name: The domain you want to get the progress information about.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"changeid": The specific change ID for which you want to get progress information. This is an optional parameter. If omitted, the service returns information about the most recent configuration change.

Main.Elasticsearch_Service.describe_elasticsearch_domain — Method

describe_elasticsearch_domain(domain_name)
describe_elasticsearch_domain(domain_name, params::Dict{String,<:Any})

Returns domain configuration information about the specified Elasticsearch domain, including the domain ID, domain endpoint, and domain ARN.

Arguments

domain_name: The name of the Elasticsearch domain for which you want information.

Main.Elasticsearch_Service.describe_elasticsearch_domain_config — Method

describe_elasticsearch_domain_config(domain_name)
describe_elasticsearch_domain_config(domain_name, params::Dict{String,<:Any})

Provides cluster configuration information about the specified Elasticsearch domain, such as the state, creation date, update version, and update date for cluster options.

Arguments

domain_name: The Elasticsearch domain that you want to get information about.

Main.Elasticsearch_Service.describe_elasticsearch_domains — Method

describe_elasticsearch_domains(domain_names)
describe_elasticsearch_domains(domain_names, params::Dict{String,<:Any})

Returns domain configuration information about the specified Elasticsearch domains, including the domain ID, domain endpoint, and domain ARN.

Arguments

domain_names: The Elasticsearch domains for which you want information.

Main.Elasticsearch_Service.describe_elasticsearch_instance_type_limits — Method

describe_elasticsearch_instance_type_limits(elasticsearch_version, instance_type)
describe_elasticsearch_instance_type_limits(elasticsearch_version, instance_type, params::Dict{String,<:Any})

Describe Elasticsearch Limits for a given InstanceType and ElasticsearchVersion. When modifying existing Domain, specify the DomainName to know what Limits are supported for modifying.

Arguments

elasticsearch_version: Version of Elasticsearch for which Limits are needed.

instance_type: The instance type for an Elasticsearch cluster for which Elasticsearch Limits are needed.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"domainName": DomainName represents the name of the Domain that we are trying to modify. This should be present only if we are querying for Elasticsearch Limits for existing domain.

Main.Elasticsearch_Service.describe_inbound_cross_cluster_search_connections — Method

describe_inbound_cross_cluster_search_connections()
describe_inbound_cross_cluster_search_connections(params::Dict{String,<:Any})

Lists all the inbound cross-cluster search connections for a destination domain.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Filters": A list of filters used to match properties for inbound cross-cluster search connection. Available Filter names for this operation are: cross-cluster-search-connection-id source-domain-info.domain-name source-domain-info.owner-id source-domain-info.region destination-domain-info.domain-name

"MaxResults": Set this value to limit the number of results returned. If not specified, defaults to 100.

"NextToken": NextToken is sent in case the earlier API call results contain the NextToken. It is used for pagination.

Main.Elasticsearch_Service.describe_outbound_cross_cluster_search_connections — Method

describe_outbound_cross_cluster_search_connections()
describe_outbound_cross_cluster_search_connections(params::Dict{String,<:Any})

Lists all the outbound cross-cluster search connections for a source domain.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Filters": A list of filters used to match properties for outbound cross-cluster search connection. Available Filter names for this operation are: cross-cluster-search-connection-id destination-domain-info.domain-name destination-domain-info.owner-id destination-domain-info.region source-domain-info.domain-name

"MaxResults": Set this value to limit the number of results returned. If not specified, defaults to 100.

"NextToken": NextToken is sent in case the earlier API call results contain the NextToken. It is used for pagination.

Main.Elasticsearch_Service.describe_packages — Method

describe_packages()
describe_packages(params::Dict{String,<:Any})

Describes all packages available to Amazon ES. Includes options for filtering, limiting the number of results, and pagination.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Filters": Only returns packages that match the DescribePackagesFilterList values.

"MaxResults": Limits results to a maximum number of packages.

"NextToken": Used for pagination. Only necessary if a previous API call includes a non-null NextToken value. If provided, returns results for the next page.

Main.Elasticsearch_Service.describe_reserved_elasticsearch_instance_offerings — Method

describe_reserved_elasticsearch_instance_offerings()
describe_reserved_elasticsearch_instance_offerings(params::Dict{String,<:Any})

Lists available reserved Elasticsearch instance offerings.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Set this value to limit the number of results returned. If not specified, defaults to 100.

"nextToken": NextToken should be sent in case if earlier API call produced result containing NextToken. It is used for pagination.

"offeringId": The offering identifier filter value. Use this parameter to show only the available offering that matches the specified reservation identifier.

Main.Elasticsearch_Service.describe_reserved_elasticsearch_instances — Method

describe_reserved_elasticsearch_instances()
describe_reserved_elasticsearch_instances(params::Dict{String,<:Any})

Returns information about reserved Elasticsearch instances for this account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Set this value to limit the number of results returned. If not specified, defaults to 100.

"nextToken": NextToken should be sent in case if earlier API call produced result containing NextToken. It is used for pagination.

"reservationId": The reserved instance identifier filter value. Use this parameter to show only the reservation that matches the specified reserved Elasticsearch instance ID.

Main.Elasticsearch_Service.describe_vpc_endpoints — Method

describe_vpc_endpoints(vpc_endpoint_ids)
describe_vpc_endpoints(vpc_endpoint_ids, params::Dict{String,<:Any})

Describes one or more Amazon OpenSearch Service-managed VPC endpoints.

Arguments

vpc_endpoint_ids: The unique identifiers of the endpoints to get information about.

Main.Elasticsearch_Service.dissociate_package — Method

dissociate_package(domain_name, package_id)
dissociate_package(domain_name, package_id, params::Dict{String,<:Any})

Dissociates a package from the Amazon ES domain.

Arguments

domain_name: Name of the domain that you want to associate the package with.

package_id: Internal ID of the package that you want to associate with a domain. Use DescribePackages to find this value.

Main.Elasticsearch_Service.get_compatible_elasticsearch_versions — Method

get_compatible_elasticsearch_versions()
get_compatible_elasticsearch_versions(params::Dict{String,<:Any})

Returns a list of upgrade compatible Elastisearch versions. You can optionally pass a DomainName to get all upgrade compatible Elasticsearch versions for that specific domain.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"domainName":

Main.Elasticsearch_Service.get_package_version_history — Method

get_package_version_history(package_id)
get_package_version_history(package_id, params::Dict{String,<:Any})

Returns a list of versions of the package, along with their creation time and commit message.

Arguments

package_id: Returns an audit history of versions of the package.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Limits results to a maximum number of versions.

"nextToken": Used for pagination. Only necessary if a previous API call includes a non-null NextToken value. If provided, returns results for the next page.

Main.Elasticsearch_Service.get_upgrade_history — Method

get_upgrade_history(domain_name)
get_upgrade_history(domain_name, params::Dict{String,<:Any})

Retrieves the complete history of the last 10 upgrades that were performed on the domain.

Arguments

domain_name:

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults":

"nextToken":

Main.Elasticsearch_Service.get_upgrade_status — Method

get_upgrade_status(domain_name)
get_upgrade_status(domain_name, params::Dict{String,<:Any})

Retrieves the latest status of the last upgrade or upgrade eligibility check that was performed on the domain.

Arguments

domain_name:

Main.Elasticsearch_Service.list_domain_names — Method

list_domain_names()
list_domain_names(params::Dict{String,<:Any})

Returns the name of all Elasticsearch domains owned by the current user's account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"engineType": Optional parameter to filter the output by domain engine type. Acceptable values are 'Elasticsearch' and 'OpenSearch'.

Main.Elasticsearch_Service.list_domains_for_package — Method

list_domains_for_package(package_id)
list_domains_for_package(package_id, params::Dict{String,<:Any})

Lists all Amazon ES domains associated with the package.

Arguments

package_id: The package for which to list domains.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Limits results to a maximum number of domains.

"nextToken": Used for pagination. Only necessary if a previous API call includes a non-null NextToken value. If provided, returns results for the next page.

Main.Elasticsearch_Service.list_elasticsearch_instance_types — Method

list_elasticsearch_instance_types(elasticsearch_version)
list_elasticsearch_instance_types(elasticsearch_version, params::Dict{String,<:Any})

List all Elasticsearch instance types that are supported for given ElasticsearchVersion

Arguments

elasticsearch_version: Version of Elasticsearch for which list of supported elasticsearch instance types are needed.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"domainName": DomainName represents the name of the Domain that we are trying to modify. This should be present only if we are querying for list of available Elasticsearch instance types when modifying existing domain.

"maxResults": Set this value to limit the number of results returned. Value provided must be greater than 30 else it wont be honored.

"nextToken": NextToken should be sent in case if earlier API call produced result containing NextToken. It is used for pagination.

Main.Elasticsearch_Service.list_elasticsearch_versions — Method

list_elasticsearch_versions()
list_elasticsearch_versions(params::Dict{String,<:Any})

List all supported Elasticsearch versions

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Set this value to limit the number of results returned. Value provided must be greater than 10 else it wont be honored.

"nextToken":

Main.Elasticsearch_Service.list_packages_for_domain — Method

list_packages_for_domain(domain_name)
list_packages_for_domain(domain_name, params::Dict{String,<:Any})

Lists all packages associated with the Amazon ES domain.

Arguments

domain_name: The name of the domain for which you want to list associated packages.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Limits results to a maximum number of packages.

"nextToken": Used for pagination. Only necessary if a previous API call includes a non-null NextToken value. If provided, returns results for the next page.

Main.Elasticsearch_Service.list_tags — Method

list_tags(arn)
list_tags(arn, params::Dict{String,<:Any})

Returns all tags for the given Elasticsearch domain.

Arguments

arn: Specify the ARN for the Elasticsearch domain to which the tags are attached that you want to view.

Main.Elasticsearch_Service.list_vpc_endpoint_access — Method

list_vpc_endpoint_access(domain_name)
list_vpc_endpoint_access(domain_name, params::Dict{String,<:Any})

Retrieves information about each principal that is allowed to access a given Amazon OpenSearch Service domain through the use of an interface VPC endpoint.

Arguments

domain_name: The name of the OpenSearch Service domain to retrieve access information for.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nextToken": Provides an identifier to allow retrieval of paginated results.

Main.Elasticsearch_Service.list_vpc_endpoints — Method

list_vpc_endpoints()
list_vpc_endpoints(params::Dict{String,<:Any})

Retrieves all Amazon OpenSearch Service-managed VPC endpoints in the current account and Region.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nextToken": Identifier to allow retrieval of paginated results.

Main.Elasticsearch_Service.list_vpc_endpoints_for_domain — Method

list_vpc_endpoints_for_domain(domain_name)
list_vpc_endpoints_for_domain(domain_name, params::Dict{String,<:Any})

Retrieves all Amazon OpenSearch Service-managed VPC endpoints associated with a particular domain.

Arguments

domain_name: Name of the ElasticSearch domain whose VPC endpoints are to be listed.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nextToken": Provides an identifier to allow retrieval of paginated results.

Main.Elasticsearch_Service.purchase_reserved_elasticsearch_instance_offering — Method

purchase_reserved_elasticsearch_instance_offering(reservation_name, reserved_elasticsearch_instance_offering_id)
purchase_reserved_elasticsearch_instance_offering(reservation_name, reserved_elasticsearch_instance_offering_id, params::Dict{String,<:Any})

Allows you to purchase reserved Elasticsearch instances.

Arguments

reservation_name: A customer-specified identifier to track this reservation.

reserved_elasticsearch_instance_offering_id: The ID of the reserved Elasticsearch instance offering to purchase.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"InstanceCount": The number of Elasticsearch instances to reserve.

Main.Elasticsearch_Service.reject_inbound_cross_cluster_search_connection — Method

reject_inbound_cross_cluster_search_connection(connection_id)
reject_inbound_cross_cluster_search_connection(connection_id, params::Dict{String,<:Any})

Allows the destination domain owner to reject an inbound cross-cluster search connection request.

Arguments

connection_id: The id of the inbound connection that you want to reject.

Main.Elasticsearch_Service.remove_tags — Method

remove_tags(arn, tag_keys)
remove_tags(arn, tag_keys, params::Dict{String,<:Any})

Removes the specified set of tags from the specified Elasticsearch domain.

Arguments

arn: Specifies the ARN for the Elasticsearch domain from which you want to delete the specified tags.

tag_keys: Specifies the TagKey list which you want to remove from the Elasticsearch domain.

Main.Elasticsearch_Service.revoke_vpc_endpoint_access — Method

revoke_vpc_endpoint_access(account, domain_name)
revoke_vpc_endpoint_access(account, domain_name, params::Dict{String,<:Any})

Revokes access to an Amazon OpenSearch Service domain that was provided through an interface VPC endpoint.

Arguments

account: The account ID to revoke access from.

domain_name: The name of the OpenSearch Service domain.

Main.Elasticsearch_Service.start_elasticsearch_service_software_update — Method

start_elasticsearch_service_software_update(domain_name)
start_elasticsearch_service_software_update(domain_name, params::Dict{String,<:Any})

Schedules a service software update for an Amazon ES domain.

Arguments

domain_name: The name of the domain that you want to update to the latest service software.

Main.Elasticsearch_Service.update_elasticsearch_domain_config — Method

update_elasticsearch_domain_config(domain_name)
update_elasticsearch_domain_config(domain_name, params::Dict{String,<:Any})

Modifies the cluster configuration of the specified Elasticsearch domain, setting as setting the instance type and the number of instances.

Arguments

domain_name: The name of the Elasticsearch domain that you are updating.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AccessPolicies": IAM access policy as a JSON-formatted string.

"AdvancedOptions": Modifies the advanced option to allow references to indices in an HTTP request body. Must be false when configuring access to individual sub-resources. By default, the value is true. See Configuration Advanced Options for more information.

"AdvancedSecurityOptions": Specifies advanced security options.

"AutoTuneOptions": Specifies Auto-Tune options.

"CognitoOptions": Options to specify the Cognito user and identity pools for Kibana authentication. For more information, see Amazon Cognito Authentication for Kibana.

"DomainEndpointOptions": Options to specify configuration that will be applied to the domain endpoint.

"DryRun": This flag, when set to True, specifies whether the UpdateElasticsearchDomain request should return the results of validation checks without actually applying the change. This flag, when set to True, specifies the deployment mechanism through which the update shall be applied on the domain. This will not actually perform the Update.

"EBSOptions": Specify the type and size of the EBS volume that you want to use.

"ElasticsearchClusterConfig": The type and number of instances to instantiate for the domain cluster.

"EncryptionAtRestOptions": Specifies the Encryption At Rest Options.

"LogPublishingOptions": Map of LogType and LogPublishingOption, each containing options to publish a given type of Elasticsearch log.

"NodeToNodeEncryptionOptions": Specifies the NodeToNodeEncryptionOptions.

"SnapshotOptions": Option to set the time, in UTC format, for the daily automated snapshot. Default value is 0 hours.

"VPCOptions": Options to specify the subnets and security groups for VPC endpoint. For more information, see Creating a VPC in VPC Endpoints for Amazon Elasticsearch Service Domains

Main.Elasticsearch_Service.update_package — Method

update_package(package_id, package_source)
update_package(package_id, package_source, params::Dict{String,<:Any})

Updates a package for use with Amazon ES domains.

Arguments

package_id: Unique identifier for the package.

package_source:

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"CommitMessage": An info message for the new version which will be shown as part of GetPackageVersionHistoryResponse.

"PackageDescription": New description of the package.

Main.Elasticsearch_Service.update_vpc_endpoint — Method

update_vpc_endpoint(vpc_endpoint_id, vpc_options)
update_vpc_endpoint(vpc_endpoint_id, vpc_options, params::Dict{String,<:Any})

Modifies an Amazon OpenSearch Service-managed interface VPC endpoint.

Arguments

vpc_endpoint_id: Unique identifier of the VPC endpoint to be updated.

vpc_options: The security groups and/or subnets to add, remove, or modify.

Main.Elasticsearch_Service.upgrade_elasticsearch_domain — Method

upgrade_elasticsearch_domain(domain_name, target_version)
upgrade_elasticsearch_domain(domain_name, target_version, params::Dict{String,<:Any})

Allows you to either upgrade your domain or perform an Upgrade eligibility check to a compatible Elasticsearch version.

Arguments

domain_name:

target_version: The version of Elasticsearch that you intend to upgrade the domain to.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"PerformCheckOnly": This flag, when set to True, indicates that an Upgrade Eligibility Check needs to be performed. This will not actually perform the Upgrade.