Emr Containers

This page documents function available when using the Emr_Containers module, created with @service Emr_Containers.

Index

Main.Emr_Containers.cancel_job_run

Main.Emr_Containers.create_managed_endpoint

Main.Emr_Containers.create_virtual_cluster

Main.Emr_Containers.delete_managed_endpoint

Main.Emr_Containers.delete_virtual_cluster

Main.Emr_Containers.describe_job_run

Main.Emr_Containers.describe_managed_endpoint

Main.Emr_Containers.describe_virtual_cluster

Main.Emr_Containers.list_job_runs

Main.Emr_Containers.list_managed_endpoints

Main.Emr_Containers.list_tags_for_resource

Main.Emr_Containers.list_virtual_clusters

Main.Emr_Containers.start_job_run

Main.Emr_Containers.tag_resource

Main.Emr_Containers.untag_resource

Documentation

Main.Emr_Containers.cancel_job_run — Method

cancel_job_run(job_run_id, virtual_cluster_id)
cancel_job_run(job_run_id, virtual_cluster_id, params::Dict{String,<:Any})

Cancels a job run. A job run is a unit of work, such as a Spark jar, PySpark script, or SparkSQL query, that you submit to Amazon EMR on EKS.

Arguments

job_run_id: The ID of the job run to cancel.

virtual_cluster_id: The ID of the virtual cluster for which the job run will be canceled.

Main.Emr_Containers.create_managed_endpoint — Method

create_managed_endpoint(client_token, execution_role_arn, name, release_label, type, virtual_cluster_id)
create_managed_endpoint(client_token, execution_role_arn, name, release_label, type, virtual_cluster_id, params::Dict{String,<:Any})

Creates a managed endpoint. A managed endpoint is a gateway that connects EMR Studio to Amazon EMR on EKS so that EMR Studio can communicate with your virtual cluster.

Arguments

client_token: The client idempotency token for this create call.

execution_role_arn: The ARN of the execution role.

name: The name of the managed endpoint.

release_label: The Amazon EMR release version.

type: The type of the managed endpoint.

virtual_cluster_id: The ID of the virtual cluster for which a managed endpoint is created.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"certificateArn": The certificate ARN provided by users for the managed endpoint. This fiedd is under deprecation and will be removed in future releases.

"configurationOverrides": The configuration settings that will be used to override existing configurations.

"tags": The tags of the managed endpoint.

Main.Emr_Containers.create_virtual_cluster — Method

create_virtual_cluster(client_token, container_provider, name)
create_virtual_cluster(client_token, container_provider, name, params::Dict{String,<:Any})

Creates a virtual cluster. Virtual cluster is a managed entity on Amazon EMR on EKS. You can create, describe, list and delete virtual clusters. They do not consume any additional resource in your system. A single virtual cluster maps to a single Kubernetes namespace. Given this relationship, you can model virtual clusters the same way you model Kubernetes namespaces to meet your requirements.

Arguments

client_token: The client token of the virtual cluster.

container_provider: The container provider of the virtual cluster.

name: The specified name of the virtual cluster.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"tags": The tags assigned to the virtual cluster.

Main.Emr_Containers.delete_managed_endpoint — Method

delete_managed_endpoint(endpoint_id, virtual_cluster_id)
delete_managed_endpoint(endpoint_id, virtual_cluster_id, params::Dict{String,<:Any})

Deletes a managed endpoint. A managed endpoint is a gateway that connects EMR Studio to Amazon EMR on EKS so that EMR Studio can communicate with your virtual cluster.

Arguments

endpoint_id: The ID of the managed endpoint.

virtual_cluster_id: The ID of the endpoint's virtual cluster.

Main.Emr_Containers.delete_virtual_cluster — Method

delete_virtual_cluster(virtual_cluster_id)
delete_virtual_cluster(virtual_cluster_id, params::Dict{String,<:Any})

Deletes a virtual cluster. Virtual cluster is a managed entity on Amazon EMR on EKS. You can create, describe, list and delete virtual clusters. They do not consume any additional resource in your system. A single virtual cluster maps to a single Kubernetes namespace. Given this relationship, you can model virtual clusters the same way you model Kubernetes namespaces to meet your requirements.

Arguments

virtual_cluster_id: The ID of the virtual cluster that will be deleted.

Main.Emr_Containers.describe_job_run — Method

describe_job_run(job_run_id, virtual_cluster_id)
describe_job_run(job_run_id, virtual_cluster_id, params::Dict{String,<:Any})

Displays detailed information about a job run. A job run is a unit of work, such as a Spark jar, PySpark script, or SparkSQL query, that you submit to Amazon EMR on EKS.

Arguments

job_run_id: The ID of the job run request.

virtual_cluster_id: The ID of the virtual cluster for which the job run is submitted.

Main.Emr_Containers.describe_managed_endpoint — Method

describe_managed_endpoint(endpoint_id, virtual_cluster_id)
describe_managed_endpoint(endpoint_id, virtual_cluster_id, params::Dict{String,<:Any})

Displays detailed information about a managed endpoint. A managed endpoint is a gateway that connects EMR Studio to Amazon EMR on EKS so that EMR Studio can communicate with your virtual cluster.

Arguments

endpoint_id: This output displays ID of the managed endpoint.

virtual_cluster_id: The ID of the endpoint's virtual cluster.

Main.Emr_Containers.describe_virtual_cluster — Method

describe_virtual_cluster(virtual_cluster_id)
describe_virtual_cluster(virtual_cluster_id, params::Dict{String,<:Any})

Displays detailed information about a specified virtual cluster. Virtual cluster is a managed entity on Amazon EMR on EKS. You can create, describe, list and delete virtual clusters. They do not consume any additional resource in your system. A single virtual cluster maps to a single Kubernetes namespace. Given this relationship, you can model virtual clusters the same way you model Kubernetes namespaces to meet your requirements.

Arguments

virtual_cluster_id: The ID of the virtual cluster that will be described.

Main.Emr_Containers.list_job_runs — Method

list_job_runs(virtual_cluster_id)
list_job_runs(virtual_cluster_id, params::Dict{String,<:Any})

Lists job runs based on a set of parameters. A job run is a unit of work, such as a Spark jar, PySpark script, or SparkSQL query, that you submit to Amazon EMR on EKS.

Arguments

virtual_cluster_id: The ID of the virtual cluster for which to list the job run.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"createdAfter": The date and time after which the job runs were submitted.

"createdBefore": The date and time before which the job runs were submitted.

"maxResults": The maximum number of job runs that can be listed.

"name": The name of the job run.

"nextToken": The token for the next set of job runs to return.

"states": The states of the job run.

Main.Emr_Containers.list_managed_endpoints — Method

list_managed_endpoints(virtual_cluster_id)
list_managed_endpoints(virtual_cluster_id, params::Dict{String,<:Any})

Lists managed endpoints based on a set of parameters. A managed endpoint is a gateway that connects EMR Studio to Amazon EMR on EKS so that EMR Studio can communicate with your virtual cluster.

Arguments

virtual_cluster_id: The ID of the virtual cluster.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"createdAfter": The date and time after which the endpoints are created.

"createdBefore": The date and time before which the endpoints are created.

"maxResults": The maximum number of managed endpoints that can be listed.

"nextToken": The token for the next set of managed endpoints to return.

"states": The states of the managed endpoints.

"types": The types of the managed endpoints.

Main.Emr_Containers.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists the tags assigned to the resources.

Arguments

resource_arn: The ARN of tagged resources.

Main.Emr_Containers.list_virtual_clusters — Method

list_virtual_clusters()
list_virtual_clusters(params::Dict{String,<:Any})

Lists information about the specified virtual cluster. Virtual cluster is a managed entity on Amazon EMR on EKS. You can create, describe, list and delete virtual clusters. They do not consume any additional resource in your system. A single virtual cluster maps to a single Kubernetes namespace. Given this relationship, you can model virtual clusters the same way you model Kubernetes namespaces to meet your requirements.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"containerProviderId": The container provider ID of the virtual cluster.

"containerProviderType": The container provider type of the virtual cluster. EKS is the only supported type as of now.

"createdAfter": The date and time after which the virtual clusters are created.

"createdBefore": The date and time before which the virtual clusters are created.

"maxResults": The maximum number of virtual clusters that can be listed.

"nextToken": The token for the next set of virtual clusters to return.

"states": The states of the requested virtual clusters.

Main.Emr_Containers.start_job_run — Method

start_job_run(client_token, execution_role_arn, job_driver, release_label, virtual_cluster_id)
start_job_run(client_token, execution_role_arn, job_driver, release_label, virtual_cluster_id, params::Dict{String,<:Any})

Starts a job run. A job run is a unit of work, such as a Spark jar, PySpark script, or SparkSQL query, that you submit to Amazon EMR on EKS.

Arguments

client_token: The client idempotency token of the job run request.

execution_role_arn: The execution role ARN for the job run.

job_driver: The job driver for the job run.

release_label: The Amazon EMR release version to use for the job run.

virtual_cluster_id: The virtual cluster ID for which the job run request is submitted.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"configurationOverrides": The configuration overrides for the job run.

"name": The name of the job run.

"tags": The tags assigned to job runs.

Main.Emr_Containers.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Assigns tags to resources. A tag is a label that you assign to an AWS resource. Each tag consists of a key and an optional value, both of which you define. Tags enable you to categorize your AWS resources by attributes such as purpose, owner, or environment. When you have many resources of the same type, you can quickly identify a specific resource based on the tags you've assigned to it. For example, you can define a set of tags for your Amazon EMR on EKS clusters to help you track each cluster's owner and stack level. We recommend that you devise a consistent set of tag keys for each resource type. You can then search and filter the resources based on the tags that you add.

Arguments

resource_arn: The ARN of resources.

tags: The tags assigned to resources.

Main.Emr_Containers.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes tags from resources.

Arguments

resource_arn: The ARN of resources.

tag_keys: The tag keys of the resources.