Frauddetector

This page documents function available when using the Frauddetector module, created with @service Frauddetector.

Index

Main.Frauddetector.batch_create_variable

Main.Frauddetector.batch_get_variable

Main.Frauddetector.create_detector_version

Main.Frauddetector.create_model

Main.Frauddetector.create_model_version

Main.Frauddetector.create_rule

Main.Frauddetector.create_variable

Main.Frauddetector.delete_detector

Main.Frauddetector.delete_detector_version

Main.Frauddetector.delete_entity_type

Main.Frauddetector.delete_event

Main.Frauddetector.delete_event_type

Main.Frauddetector.delete_external_model

Main.Frauddetector.delete_label

Main.Frauddetector.delete_model

Main.Frauddetector.delete_model_version

Main.Frauddetector.delete_outcome

Main.Frauddetector.delete_rule

Main.Frauddetector.delete_variable

Main.Frauddetector.describe_detector

Main.Frauddetector.describe_model_versions

Main.Frauddetector.get_detector_version

Main.Frauddetector.get_detectors

Main.Frauddetector.get_entity_types

Main.Frauddetector.get_event_prediction

Main.Frauddetector.get_event_types

Main.Frauddetector.get_external_models

Main.Frauddetector.get_kmsencryption_key

Main.Frauddetector.get_labels

Main.Frauddetector.get_model_version

Main.Frauddetector.get_models

Main.Frauddetector.get_outcomes

Main.Frauddetector.get_rules

Main.Frauddetector.get_variables

Main.Frauddetector.list_tags_for_resource

Main.Frauddetector.put_detector

Main.Frauddetector.put_entity_type

Main.Frauddetector.put_event_type

Main.Frauddetector.put_external_model

Main.Frauddetector.put_kmsencryption_key

Main.Frauddetector.put_label

Main.Frauddetector.put_outcome

Main.Frauddetector.tag_resource

Main.Frauddetector.untag_resource

Main.Frauddetector.update_detector_version

Main.Frauddetector.update_detector_version_metadata

Main.Frauddetector.update_detector_version_status

Main.Frauddetector.update_model

Main.Frauddetector.update_model_version

Main.Frauddetector.update_model_version_status

Main.Frauddetector.update_rule_metadata

Main.Frauddetector.update_rule_version

Main.Frauddetector.update_variable

Documentation

Main.Frauddetector.batch_create_variable — Method

BatchCreateVariable()

Creates a batch of variables.

Required Parameters

variableEntries: The list of variables for the batch create variable request.

Optional Parameters

tags: A collection of key and value pairs.

Main.Frauddetector.batch_get_variable — Method

BatchGetVariable()

Gets a batch of variables.

Required Parameters

names: The list of variable names to get.

Main.Frauddetector.create_detector_version — Method

CreateDetectorVersion()

Creates a detector version. The detector version starts in a DRAFT status.

Required Parameters

detectorId: The ID of the detector under which you want to create a new version.

rules: The rules to include in the detector version.

Optional Parameters

description: The description of the detector version.

externalModelEndpoints: The Amazon Sagemaker model endpoints to include in the detector version.

modelVersions: The model versions to include in the detector version.

ruleExecutionMode: The rule execution mode for the rules included in the detector version. You can define and edit the rule mode at the detector version level, when it is in draft status. If you specify FIRSTMATCHED, Amazon Fraud Detector evaluates rules sequentially, first to last, stopping at the first matched rule. Amazon Fraud dectector then provides the outcomes for that single rule. If you specifiy ALLMATCHED, Amazon Fraud Detector evaluates all rules and returns the outcomes for all matched rules. The default behavior is FIRST_MATCHED.

tags: A collection of key and value pairs.

Main.Frauddetector.create_model — Method

CreateModel()

Creates a model using the specified model type.

Required Parameters

eventTypeName: The name of the event type.

modelId: The model ID.

modelType: The model type.

Optional Parameters

description: The model description.

tags: A collection of key and value pairs.

Main.Frauddetector.create_model_version — Method

CreateModelVersion()

Creates a version of the model using the specified model type and model id.

Required Parameters

modelId: The model ID.

modelType: The model type.

trainingDataSchema: The training data schema.

trainingDataSource: The training data source location in Amazon S3.

Optional Parameters

externalEventsDetail: Details for the external events data used for model version training. Required if trainingDataSource is EXTERNAL_EVENTS.

tags: A collection of key and value pairs.

Main.Frauddetector.create_rule — Method

CreateRule()

Creates a rule for use with the specified detector.

Required Parameters

detectorId: The detector ID for the rule's parent detector.

expression: The rule expression.

language: The language of the rule.

outcomes: The outcome or outcomes returned when the rule expression matches.

ruleId: The rule ID.

Optional Parameters

description: The rule description.

tags: A collection of key and value pairs.

Main.Frauddetector.create_variable — Method

CreateVariable()

Creates a variable.

Required Parameters

dataSource: The source of the data.

dataType: The data type.

defaultValue: The default value for the variable when no value is received.

name: The name of the variable.

Optional Parameters

description: The description.

tags: A collection of key and value pairs.

variableType: The variable type. For more information see Variable types. Valid Values: AUTHCODE | AVS | BILLINGADDRESSL1 | BILLINGADDRESSL2 | BILLINGCITY | BILLINGCOUNTRY | BILLINGNAME | BILLINGPHONE | BILLINGSTATE | BILLINGZIP | CARDBIN | CATEGORICAL | CURRENCYCODE | EMAILADDRESS | FINGERPRINT | FRAUDLABEL | FREEFORMTEXT | IPADDRESS | NUMERIC | ORDERID | PAYMENTTYPE | PHONENUMBER | PRICE | PRODUCTCATEGORY | SHIPPINGADDRESSL1 | SHIPPINGADDRESSL2 | SHIPPINGCITY | SHIPPINGCOUNTRY | SHIPPINGNAME | SHIPPINGPHONE | SHIPPINGSTATE | SHIPPINGZIP | USERAGENT

Main.Frauddetector.delete_detector — Method

DeleteDetector()

Deletes the detector. Before deleting a detector, you must first delete all detector versions and rule versions associated with the detector. When you delete a detector, Amazon Fraud Detector permanently deletes the detector and the data is no longer stored in Amazon Fraud Detector.

Required Parameters

detectorId: The ID of the detector to delete.

Main.Frauddetector.delete_detector_version — Method

DeleteDetectorVersion()

Deletes the detector version. You cannot delete detector versions that are in ACTIVE status. When you delete a detector version, Amazon Fraud Detector permanently deletes the detector and the data is no longer stored in Amazon Fraud Detector.

Required Parameters

detectorId: The ID of the parent detector for the detector version to delete.

detectorVersionId: The ID of the detector version to delete.

Main.Frauddetector.delete_entity_type — Method

DeleteEntityType()

Deletes an entity type. You cannot delete an entity type that is included in an event type. When you delete an entity type, Amazon Fraud Detector permanently deletes that entity type and the data is no longer stored in Amazon Fraud Detector.

Required Parameters

name: The name of the entity type to delete.

Main.Frauddetector.delete_event — Method

DeleteEvent()

Deletes the specified event. When you delete an event, Amazon Fraud Detector permanently deletes that event and the event data is no longer stored in Amazon Fraud Detector.

Required Parameters

eventId: The ID of the event to delete.

eventTypeName: The name of the event type.

Main.Frauddetector.delete_event_type — Method

DeleteEventType()

Deletes an event type. You cannot delete an event type that is used in a detector or a model. When you delete an entity type, Amazon Fraud Detector permanently deletes that entity type and the data is no longer stored in Amazon Fraud Detector.

Required Parameters

name: The name of the event type to delete.

Main.Frauddetector.delete_external_model — Method

DeleteExternalModel()

Removes a SageMaker model from Amazon Fraud Detector. You can remove an Amazon SageMaker model if it is not associated with a detector version. Removing a SageMaker model disconnects it from Amazon Fraud Detector, but the model remains available in SageMaker.

Required Parameters

modelEndpoint: The endpoint of the Amazon Sagemaker model to delete.

Main.Frauddetector.delete_label — Method

DeleteLabel()

Deletes a label. You cannot delete labels that are included in an event type in Amazon Fraud Detector. You cannot delete a label assigned to an event ID. You must first delete the relevant event ID. When you delete a label, Amazon Fraud Detector permanently deletes that label and the data is no longer stored in Amazon Fraud Detector.

Required Parameters

name: The name of the label to delete.

Main.Frauddetector.delete_model — Method

DeleteModel()

Deletes a model. You can delete models and model versions in Amazon Fraud Detector, provided that they are not associated with a detector version. When you delete a model, Amazon Fraud Detector permanently deletes that model and the data is no longer stored in Amazon Fraud Detector.

Required Parameters

modelId: The model ID of the model to delete.

modelType: The model type of the model to delete.

Main.Frauddetector.delete_model_version — Method

DeleteModelVersion()

Deletes a model version. You can delete models and model versions in Amazon Fraud Detector, provided that they are not associated with a detector version. When you delete a model version, Amazon Fraud Detector permanently deletes that model version and the data is no longer stored in Amazon Fraud Detector.

Required Parameters

modelId: The model ID of the model version to delete.

modelType: The model type of the model version to delete.

modelVersionNumber: The model version number of the model version to delete.

Main.Frauddetector.delete_outcome — Method

DeleteOutcome()

Deletes an outcome. You cannot delete an outcome that is used in a rule version. When you delete an outcome, Amazon Fraud Detector permanently deletes that outcome and the data is no longer stored in Amazon Fraud Detector.

Required Parameters

name: The name of the outcome to delete.

Main.Frauddetector.delete_rule — Method

DeleteRule()

Deletes the rule. You cannot delete a rule if it is used by an ACTIVE or INACTIVE detector version. When you delete a rule, Amazon Fraud Detector permanently deletes that rule and the data is no longer stored in Amazon Fraud Detector.

Required Parameters

rule:

Main.Frauddetector.delete_variable — Method

DeleteVariable()

Deletes a variable. You can't delete variables that are included in an event type in Amazon Fraud Detector. Amazon Fraud Detector automatically deletes model output variables and SageMaker model output variables when you delete the model. You can't delete these variables manually. When you delete a variable, Amazon Fraud Detector permanently deletes that variable and the data is no longer stored in Amazon Fraud Detector.

Required Parameters

name: The name of the variable to delete.

Main.Frauddetector.describe_detector — Method

DescribeDetector()

Gets all versions for a specified detector.

Required Parameters

detectorId: The detector ID.

Optional Parameters

maxResults: The maximum number of results to return for the request.

nextToken: The next token from the previous response.

Main.Frauddetector.describe_model_versions — Method

DescribeModelVersions()

Gets all of the model versions for the specified model type or for the specified model type and model ID. You can also get details for a single, specified model version.

Optional Parameters

maxResults: The maximum number of results to return.

modelId: The model ID.

modelType: The model type.

modelVersionNumber: The model version number.

nextToken: The next token from the previous results.

Main.Frauddetector.get_detector_version — Method

GetDetectorVersion()

Gets a particular detector version.

Required Parameters

detectorId: The detector ID.

detectorVersionId: The detector version ID.

Main.Frauddetector.get_detectors — Method

GetDetectors()

Gets all detectors or a single detector if a detectorId is specified. This is a paginated API. If you provide a null maxResults, this action retrieves a maximum of 10 records per page. If you provide a maxResults, the value must be between 5 and 10. To get the next page results, provide the pagination token from the GetDetectorsResponse as part of your request. A null pagination token fetches the records from the beginning.

Optional Parameters

detectorId: The detector ID.

maxResults: The maximum number of objects to return for the request.

nextToken: The next token for the subsequent request.

Main.Frauddetector.get_entity_types — Method

GetEntityTypes()

Gets all entity types or a specific entity type if a name is specified. This is a paginated API. If you provide a null maxResults, this action retrieves a maximum of 10 records per page. If you provide a maxResults, the value must be between 5 and 10. To get the next page results, provide the pagination token from the GetEntityTypesResponse as part of your request. A null pagination token fetches the records from the beginning.

Optional Parameters

maxResults: The maximum number of objects to return for the request.

name: The name.

nextToken: The next token for the subsequent request.

Main.Frauddetector.get_event_prediction — Method

GetEventPrediction()

Evaluates an event against a detector version. If a version ID is not provided, the detector’s (ACTIVE) version is used.

Required Parameters

detectorId: The detector ID.

entities: The entity type (associated with the detector's event type) and specific entity ID representing who performed the event. If an entity id is not available, use "UNKNOWN."

eventId: The unique ID used to identify the event.

eventTimestamp: Timestamp that defines when the event under evaluation occurred.

eventTypeName: The event type associated with the detector specified for the prediction.

eventVariables: Names of the event type's variables you defined in Amazon Fraud Detector to represent data elements and their corresponding values for the event you are sending for evaluation.

Optional Parameters

detectorVersionId: The detector version ID.

externalModelEndpointDataBlobs: The Amazon SageMaker model endpoint input data blobs.

Main.Frauddetector.get_event_types — Method

GetEventTypes()

Gets all event types or a specific event type if name is provided. This is a paginated API. If you provide a null maxResults, this action retrieves a maximum of 10 records per page. If you provide a maxResults, the value must be between 5 and 10. To get the next page results, provide the pagination token from the GetEventTypesResponse as part of your request. A null pagination token fetches the records from the beginning.

Optional Parameters

maxResults: The maximum number of objects to return for the request.

name: The name.

nextToken: The next token for the subsequent request.

Main.Frauddetector.get_external_models — Method

GetExternalModels()

Gets the details for one or more Amazon SageMaker models that have been imported into the service. This is a paginated API. If you provide a null maxResults, this actions retrieves a maximum of 10 records per page. If you provide a maxResults, the value must be between 5 and 10. To get the next page results, provide the pagination token from the GetExternalModelsResult as part of your request. A null pagination token fetches the records from the beginning.

Optional Parameters

maxResults: The maximum number of objects to return for the request.

modelEndpoint: The Amazon SageMaker model endpoint.

nextToken: The next page token for the request.

Main.Frauddetector.get_kmsencryption_key — Method

GetKMSEncryptionKey()

Gets the encryption key if a Key Management Service (KMS) customer master key (CMK) has been specified to be used to encrypt content in Amazon Fraud Detector.

Main.Frauddetector.get_labels — Method

GetLabels()

Gets all labels or a specific label if name is provided. This is a paginated API. If you provide a null maxResults, this action retrieves a maximum of 50 records per page. If you provide a maxResults, the value must be between 10 and 50. To get the next page results, provide the pagination token from the GetGetLabelsResponse as part of your request. A null pagination token fetches the records from the beginning.

Optional Parameters

maxResults: The maximum number of objects to return for the request.

name: The name of the label or labels to get.

nextToken: The next token for the subsequent request.

Main.Frauddetector.get_model_version — Method

GetModelVersion()

Gets the details of the specified model version.

Required Parameters

modelId: The model ID.

modelType: The model type.

modelVersionNumber: The model version number.

Main.Frauddetector.get_models — Method

GetModels()

Gets one or more models. Gets all models for the AWS account if no model type and no model id provided. Gets all models for the AWS account and model type, if the model type is specified but model id is not provided. Gets a specific model if (model type, model id) tuple is specified. This is a paginated API. If you provide a null maxResults, this action retrieves a maximum of 10 records per page. If you provide a maxResults, the value must be between 1 and 10. To get the next page results, provide the pagination token from the response as part of your request. A null pagination token fetches the records from the beginning.

Optional Parameters

maxResults: The maximum number of objects to return for the request.

modelId: The model ID.

modelType: The model type.

nextToken: The next token for the subsequent request.

Main.Frauddetector.get_outcomes — Method

GetOutcomes()

Gets one or more outcomes. This is a paginated API. If you provide a null maxResults, this actions retrieves a maximum of 100 records per page. If you provide a maxResults, the value must be between 50 and 100. To get the next page results, provide the pagination token from the GetOutcomesResult as part of your request. A null pagination token fetches the records from the beginning.

Optional Parameters

maxResults: The maximum number of objects to return for the request.

name: The name of the outcome or outcomes to get.

nextToken: The next page token for the request.

Main.Frauddetector.get_rules — Method

GetRules()

Get all rules for a detector (paginated) if ruleId and ruleVersion are not specified. Gets all rules for the detector and the ruleId if present (paginated). Gets a specific rule if both the ruleId and the ruleVersion are specified. This is a paginated API. Providing null maxResults results in retrieving maximum of 100 records per page. If you provide maxResults the value must be between 50 and 100. To get the next page result, a provide a pagination token from GetRulesResult as part of your request. Null pagination token fetches the records from the beginning.

Required Parameters

detectorId: The detector ID.

Optional Parameters

maxResults: The maximum number of rules to return for the request.

nextToken: The next page token.

ruleId: The rule ID.

ruleVersion: The rule version.

Main.Frauddetector.get_variables — Method

GetVariables()

Gets all of the variables or the specific variable. This is a paginated API. Providing null maxSizePerPage results in retrieving maximum of 100 records per page. If you provide maxSizePerPage the value must be between 50 and 100. To get the next page result, a provide a pagination token from GetVariablesResult as part of your request. Null pagination token fetches the records from the beginning.

Optional Parameters

maxResults: The max size per page determined for the get variable request.

name: The name of the variable.

nextToken: The next page token of the get variable request.

Main.Frauddetector.list_tags_for_resource — Method

ListTagsForResource()

Lists all tags associated with the resource. This is a paginated API. To get the next page results, provide the pagination token from the response as part of your request. A null pagination token fetches the records from the beginning.

Required Parameters

resourceARN: The ARN that specifies the resource whose tags you want to list.

Optional Parameters

maxResults: The maximum number of objects to return for the request.

nextToken: The next token from the previous results.

Main.Frauddetector.put_detector — Method

PutDetector()

Creates or updates a detector.

Required Parameters

detectorId: The detector ID.

eventTypeName: The name of the event type.

Optional Parameters

description: The description of the detector.

tags: A collection of key and value pairs.

Main.Frauddetector.put_entity_type — Method

PutEntityType()

Creates or updates an entity type. An entity represents who is performing the event. As part of a fraud prediction, you pass the entity ID to indicate the specific entity who performed the event. An entity type classifies the entity. Example classifications include customer, merchant, or account.

Required Parameters

name: The name of the entity type.

Optional Parameters

description: The description.

tags: A collection of key and value pairs.

Main.Frauddetector.put_event_type — Method

PutEventType()

Creates or updates an event type. An event is a business activity that is evaluated for fraud risk. With Amazon Fraud Detector, you generate fraud predictions for events. An event type defines the structure for an event sent to Amazon Fraud Detector. This includes the variables sent as part of the event, the entity performing the event (such as a customer), and the labels that classify the event. Example event types include online payment transactions, account registrations, and authentications.

Required Parameters

entityTypes: The entity type for the event type. Example entity types: customer, merchant, account.

eventVariables: The event type variables.

name: The name.

Optional Parameters

description: The description of the event type.

labels: The event type labels.

tags: A collection of key and value pairs.

Main.Frauddetector.put_external_model — Method

PutExternalModel()

Creates or updates an Amazon SageMaker model endpoint. You can also use this action to update the configuration of the model endpoint, including the IAM role and/or the mapped variables.

Required Parameters

inputConfiguration: The model endpoint input configuration.

invokeModelEndpointRoleArn: The IAM role used to invoke the model endpoint.

modelEndpoint: The model endpoints name.

modelEndpointStatus: The model endpoint’s status in Amazon Fraud Detector.

modelSource: The source of the model.

outputConfiguration: The model endpoint output configuration.

Optional Parameters

tags: A collection of key and value pairs.

Main.Frauddetector.put_kmsencryption_key — Method

PutKMSEncryptionKey()

Specifies the Key Management Service (KMS) customer master key (CMK) to be used to encrypt content in Amazon Fraud Detector.

Required Parameters

kmsEncryptionKeyArn: The KMS encryption key ARN.

Main.Frauddetector.put_label — Method

PutLabel()

Creates or updates label. A label classifies an event as fraudulent or legitimate. Labels are associated with event types and used to train supervised machine learning models in Amazon Fraud Detector.

Required Parameters

name: The label name.

Optional Parameters

description: The label description.

tags:

Main.Frauddetector.put_outcome — Method

PutOutcome()

Creates or updates an outcome.

Required Parameters

name: The name of the outcome.

Optional Parameters

description: The outcome description.

tags: A collection of key and value pairs.

Main.Frauddetector.tag_resource — Method

TagResource()

Assigns tags to a resource.

Required Parameters

resourceARN: The resource ARN.

tags: The tags to assign to the resource.

Main.Frauddetector.untag_resource — Method

UntagResource()

Removes tags from a resource.

Required Parameters

resourceARN: The ARN of the resource from which to remove the tag.

tagKeys: The resource ARN.

Main.Frauddetector.update_detector_version — Method

UpdateDetectorVersion()

Updates a detector version. The detector version attributes that you can update include models, external model endpoints, rules, rule execution mode, and description. You can only update a DRAFT detector version.

Required Parameters

detectorId: The parent detector ID for the detector version you want to update.

detectorVersionId: The detector version ID.

externalModelEndpoints: The Amazon SageMaker model endpoints to include in the detector version.

rules: The rules to include in the detector version.

Optional Parameters

description: The detector version description.

modelVersions: The model versions to include in the detector version.

ruleExecutionMode: The rule execution mode to add to the detector. If you specify FIRSTMATCHED, Amazon Fraud Detector evaluates rules sequentially, first to last, stopping at the first matched rule. Amazon Fraud dectector then provides the outcomes for that single rule. If you specifiy ALLMATCHED, Amazon Fraud Detector evaluates all rules and returns the outcomes for all matched rules. You can define and edit the rule mode at the detector version level, when it is in draft status. The default behavior is FIRST_MATCHED.

Main.Frauddetector.update_detector_version_metadata — Method

UpdateDetectorVersionMetadata()

Updates the detector version's description. You can update the metadata for any detector version (DRAFT, ACTIVE, or INACTIVE).

Required Parameters

description: The description.

detectorId: The detector ID.

detectorVersionId: The detector version ID.

Main.Frauddetector.update_detector_version_status — Method

UpdateDetectorVersionStatus()

Updates the detector version’s status. You can perform the following promotions or demotions using UpdateDetectorVersionStatus: DRAFT to ACTIVE, ACTIVE to INACTIVE, and INACTIVE to ACTIVE.

Required Parameters

detectorId: The detector ID.

detectorVersionId: The detector version ID.

status: The new status.

Main.Frauddetector.update_model — Method

UpdateModel()

Updates a model. You can update the description attribute using this action.

Required Parameters

modelId: The model ID.

modelType: The model type.

Optional Parameters

description: The new model description.

Main.Frauddetector.update_model_version — Method

UpdateModelVersion()

Updates a model version. Updating a model version retrains an existing model version using updated training data and produces a new minor version of the model. You can update the training data set location and data access role attributes using this action. This action creates and trains a new minor version of the model, for example version 1.01, 1.02, 1.03.

Required Parameters

majorVersionNumber: The major version number.

modelId: The model ID.

modelType: The model type.

Optional Parameters

externalEventsDetail: The event details.

tags: A collection of key and value pairs.

Main.Frauddetector.update_model_version_status — Method

UpdateModelVersionStatus()

Updates the status of a model version. You can perform the following status updates: Change the TRAINING_COMPLETE status to ACTIVE. Change ACTIVEto INACTIVE.

Required Parameters

modelId: The model ID of the model version to update.

modelType: The model type.

modelVersionNumber: The model version number.

status: The model version status.

Main.Frauddetector.update_rule_metadata — Method

UpdateRuleMetadata()

Updates a rule's metadata. The description attribute can be updated.

Required Parameters

description: The rule description.

rule: The rule to update.

Main.Frauddetector.update_rule_version — Method

UpdateRuleVersion()

Updates a rule version resulting in a new rule version. Updates a rule version resulting in a new rule version (version 1, 2, 3 ...).

Required Parameters

expression: The rule expression.

language: The language.

outcomes: The outcomes.

rule: The rule to update.

Optional Parameters

description: The description.

tags: The tags to assign to the rule version.

Main.Frauddetector.update_variable — Method

UpdateVariable()

Updates a variable.

Required Parameters

name: The name of the variable.

Optional Parameters

defaultValue: The new default value of the variable.

description: The new description.

variableType: The variable type. For more information see Variable types.