Greengrass

This page documents function available when using the Greengrass module, created with @service Greengrass.

Index

Main.Greengrass.associate_role_to_group

Main.Greengrass.associate_service_role_to_account

Main.Greengrass.create_connector_definition

Main.Greengrass.create_connector_definition_version

Main.Greengrass.create_core_definition

Main.Greengrass.create_core_definition_version

Main.Greengrass.create_deployment

Main.Greengrass.create_device_definition

Main.Greengrass.create_device_definition_version

Main.Greengrass.create_function_definition

Main.Greengrass.create_function_definition_version

Main.Greengrass.create_group

Main.Greengrass.create_group_certificate_authority

Main.Greengrass.create_group_version

Main.Greengrass.create_logger_definition

Main.Greengrass.create_logger_definition_version

Main.Greengrass.create_resource_definition

Main.Greengrass.create_resource_definition_version

Main.Greengrass.create_software_update_job

Main.Greengrass.create_subscription_definition

Main.Greengrass.create_subscription_definition_version

Main.Greengrass.delete_connector_definition

Main.Greengrass.delete_core_definition

Main.Greengrass.delete_device_definition

Main.Greengrass.delete_function_definition

Main.Greengrass.delete_group

Main.Greengrass.delete_logger_definition

Main.Greengrass.delete_resource_definition

Main.Greengrass.delete_subscription_definition

Main.Greengrass.disassociate_role_from_group

Main.Greengrass.disassociate_service_role_from_account

Main.Greengrass.get_associated_role

Main.Greengrass.get_bulk_deployment_status

Main.Greengrass.get_connectivity_info

Main.Greengrass.get_connector_definition

Main.Greengrass.get_connector_definition_version

Main.Greengrass.get_core_definition

Main.Greengrass.get_core_definition_version

Main.Greengrass.get_deployment_status

Main.Greengrass.get_device_definition

Main.Greengrass.get_device_definition_version

Main.Greengrass.get_function_definition

Main.Greengrass.get_function_definition_version

Main.Greengrass.get_group

Main.Greengrass.get_group_certificate_authority

Main.Greengrass.get_group_certificate_configuration

Main.Greengrass.get_group_version

Main.Greengrass.get_logger_definition

Main.Greengrass.get_logger_definition_version

Main.Greengrass.get_resource_definition

Main.Greengrass.get_resource_definition_version

Main.Greengrass.get_service_role_for_account

Main.Greengrass.get_subscription_definition

Main.Greengrass.get_subscription_definition_version

Main.Greengrass.get_thing_runtime_configuration

Main.Greengrass.list_bulk_deployment_detailed_reports

Main.Greengrass.list_bulk_deployments

Main.Greengrass.list_connector_definition_versions

Main.Greengrass.list_connector_definitions

Main.Greengrass.list_core_definition_versions

Main.Greengrass.list_core_definitions

Main.Greengrass.list_deployments

Main.Greengrass.list_device_definition_versions

Main.Greengrass.list_device_definitions

Main.Greengrass.list_function_definition_versions

Main.Greengrass.list_function_definitions

Main.Greengrass.list_group_certificate_authorities

Main.Greengrass.list_group_versions

Main.Greengrass.list_groups

Main.Greengrass.list_logger_definition_versions

Main.Greengrass.list_logger_definitions

Main.Greengrass.list_resource_definition_versions

Main.Greengrass.list_resource_definitions

Main.Greengrass.list_subscription_definition_versions

Main.Greengrass.list_subscription_definitions

Main.Greengrass.list_tags_for_resource

Main.Greengrass.reset_deployments

Main.Greengrass.start_bulk_deployment

Main.Greengrass.stop_bulk_deployment

Main.Greengrass.tag_resource

Main.Greengrass.untag_resource

Main.Greengrass.update_connectivity_info

Main.Greengrass.update_connector_definition

Main.Greengrass.update_core_definition

Main.Greengrass.update_device_definition

Main.Greengrass.update_function_definition

Main.Greengrass.update_group

Main.Greengrass.update_group_certificate_configuration

Main.Greengrass.update_logger_definition

Main.Greengrass.update_resource_definition

Main.Greengrass.update_subscription_definition

Main.Greengrass.update_thing_runtime_configuration

Documentation

Main.Greengrass.associate_role_to_group — Method

associate_role_to_group(group_id, role_arn)
associate_role_to_group(group_id, role_arn, params::Dict{String,<:Any})

Associates a role with a group. Your Greengrass core will use the role to access AWS cloud services. The role's permissions should allow Greengrass core Lambda functions to perform actions against the cloud.

Arguments

group_id: The ID of the Greengrass group.

role_arn: The ARN of the role you wish to associate with this group. The existence of the role is not validated.

Main.Greengrass.associate_service_role_to_account — Method

associate_service_role_to_account(role_arn)
associate_service_role_to_account(role_arn, params::Dict{String,<:Any})

Associates a role with your account. AWS IoT Greengrass will use the role to access your Lambda functions and AWS IoT resources. This is necessary for deployments to succeed. The role must have at least minimum permissions in the policy ''AWSGreengrassResourceAccessRolePolicy''.

Arguments

role_arn: The ARN of the service role you wish to associate with your account.

Main.Greengrass.create_connector_definition — Method

create_connector_definition()
create_connector_definition(params::Dict{String,<:Any})

Creates a connector definition. You may provide the initial version of the connector definition now or use ''CreateConnectorDefinitionVersion'' at a later time.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"InitialVersion": Information about the initial version of the connector definition.

"Name": The name of the connector definition.

"X-Amzn-Client-Token": A client token used to correlate requests and responses.

"tags": Tag(s) to add to the new resource.

Main.Greengrass.create_connector_definition_version — Method

create_connector_definition_version(connector_definition_id)
create_connector_definition_version(connector_definition_id, params::Dict{String,<:Any})

Creates a version of a connector definition which has already been defined.

Arguments

connector_definition_id: The ID of the connector definition.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Connectors": A list of references to connectors in this version, with their corresponding configuration settings.

"X-Amzn-Client-Token": A client token used to correlate requests and responses.

Main.Greengrass.create_core_definition — Method

create_core_definition()
create_core_definition(params::Dict{String,<:Any})

Creates a core definition. You may provide the initial version of the core definition now or use ''CreateCoreDefinitionVersion'' at a later time. Greengrass groups must each contain exactly one Greengrass core.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"InitialVersion": Information about the initial version of the core definition.

"Name": The name of the core definition.

"X-Amzn-Client-Token": A client token used to correlate requests and responses.

"tags": Tag(s) to add to the new resource.

Main.Greengrass.create_core_definition_version — Method

create_core_definition_version(core_definition_id)
create_core_definition_version(core_definition_id, params::Dict{String,<:Any})

Creates a version of a core definition that has already been defined. Greengrass groups must each contain exactly one Greengrass core.

Arguments

core_definition_id: The ID of the core definition.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Cores": A list of cores in the core definition version.

"X-Amzn-Client-Token": A client token used to correlate requests and responses.

Main.Greengrass.create_deployment — Method

create_deployment(deployment_type, group_id)
create_deployment(deployment_type, group_id, params::Dict{String,<:Any})

Creates a deployment. ''CreateDeployment'' requests are idempotent with respect to the ''X-Amzn-Client-Token'' token and the request parameters.

Arguments

deployment_type: The type of deployment. When used for ''CreateDeployment'', only ''NewDeployment'' and ''Redeployment'' are valid.

group_id: The ID of the Greengrass group.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DeploymentId": The ID of the deployment if you wish to redeploy a previous deployment.

"GroupVersionId": The ID of the group version to be deployed.

"X-Amzn-Client-Token": A client token used to correlate requests and responses.

Main.Greengrass.create_device_definition — Method

create_device_definition()
create_device_definition(params::Dict{String,<:Any})

Creates a device definition. You may provide the initial version of the device definition now or use ''CreateDeviceDefinitionVersion'' at a later time.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"InitialVersion": Information about the initial version of the device definition.

"Name": The name of the device definition.

"X-Amzn-Client-Token": A client token used to correlate requests and responses.

"tags": Tag(s) to add to the new resource.

Main.Greengrass.create_device_definition_version — Method

create_device_definition_version(device_definition_id)
create_device_definition_version(device_definition_id, params::Dict{String,<:Any})

Creates a version of a device definition that has already been defined.

Arguments

device_definition_id: The ID of the device definition.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Devices": A list of devices in the definition version.

"X-Amzn-Client-Token": A client token used to correlate requests and responses.

Main.Greengrass.create_function_definition — Method

create_function_definition()
create_function_definition(params::Dict{String,<:Any})

Creates a Lambda function definition which contains a list of Lambda functions and their configurations to be used in a group. You can create an initial version of the definition by providing a list of Lambda functions and their configurations now, or use ''CreateFunctionDefinitionVersion'' later.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"InitialVersion": Information about the initial version of the function definition.

"Name": The name of the function definition.

"X-Amzn-Client-Token": A client token used to correlate requests and responses.

"tags": Tag(s) to add to the new resource.

Main.Greengrass.create_function_definition_version — Method

create_function_definition_version(function_definition_id)
create_function_definition_version(function_definition_id, params::Dict{String,<:Any})

Creates a version of a Lambda function definition that has already been defined.

Arguments

function_definition_id: The ID of the Lambda function definition.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DefaultConfig": The default configuration that applies to all Lambda functions in this function definition version. Individual Lambda functions can override these settings.

"Functions": A list of Lambda functions in this function definition version.

"X-Amzn-Client-Token": A client token used to correlate requests and responses.

Main.Greengrass.create_group — Method

create_group(name)
create_group(name, params::Dict{String,<:Any})

Creates a group. You may provide the initial version of the group or use ''CreateGroupVersion'' at a later time. Tip: You can use the ''gggroupsetup'' package (https://github.com/awslabs/aws-greengrass-group-setup) as a library or command-line application to create and deploy Greengrass groups.

Arguments

name: The name of the group.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"InitialVersion": Information about the initial version of the group.

"X-Amzn-Client-Token": A client token used to correlate requests and responses.

"tags": Tag(s) to add to the new resource.

Main.Greengrass.create_group_certificate_authority — Method

create_group_certificate_authority(group_id)
create_group_certificate_authority(group_id, params::Dict{String,<:Any})

Creates a CA for the group. If a CA already exists, it will rotate the existing CA.

Arguments

group_id: The ID of the Greengrass group.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"X-Amzn-Client-Token": A client token used to correlate requests and responses.

Main.Greengrass.create_group_version — Method

create_group_version(group_id)
create_group_version(group_id, params::Dict{String,<:Any})

Creates a version of a group which has already been defined.

Arguments

group_id: The ID of the Greengrass group.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ConnectorDefinitionVersionArn": The ARN of the connector definition version for this group.

"CoreDefinitionVersionArn": The ARN of the core definition version for this group.

"DeviceDefinitionVersionArn": The ARN of the device definition version for this group.

"FunctionDefinitionVersionArn": The ARN of the function definition version for this group.

"LoggerDefinitionVersionArn": The ARN of the logger definition version for this group.

"ResourceDefinitionVersionArn": The ARN of the resource definition version for this group.

"SubscriptionDefinitionVersionArn": The ARN of the subscription definition version for this group.

"X-Amzn-Client-Token": A client token used to correlate requests and responses.

Main.Greengrass.create_logger_definition — Method

create_logger_definition()
create_logger_definition(params::Dict{String,<:Any})

Creates a logger definition. You may provide the initial version of the logger definition now or use ''CreateLoggerDefinitionVersion'' at a later time.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"InitialVersion": Information about the initial version of the logger definition.

"Name": The name of the logger definition.

"X-Amzn-Client-Token": A client token used to correlate requests and responses.

"tags": Tag(s) to add to the new resource.

Main.Greengrass.create_logger_definition_version — Method

create_logger_definition_version(logger_definition_id)
create_logger_definition_version(logger_definition_id, params::Dict{String,<:Any})

Creates a version of a logger definition that has already been defined.

Arguments

logger_definition_id: The ID of the logger definition.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Loggers": A list of loggers.

"X-Amzn-Client-Token": A client token used to correlate requests and responses.

Main.Greengrass.create_resource_definition — Method

create_resource_definition()
create_resource_definition(params::Dict{String,<:Any})

Creates a resource definition which contains a list of resources to be used in a group. You can create an initial version of the definition by providing a list of resources now, or use ''CreateResourceDefinitionVersion'' later.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"InitialVersion": Information about the initial version of the resource definition.

"Name": The name of the resource definition.

"X-Amzn-Client-Token": A client token used to correlate requests and responses.

"tags": Tag(s) to add to the new resource.

Main.Greengrass.create_resource_definition_version — Method

create_resource_definition_version(resource_definition_id)
create_resource_definition_version(resource_definition_id, params::Dict{String,<:Any})

Creates a version of a resource definition that has already been defined.

Arguments

resource_definition_id: The ID of the resource definition.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Resources": A list of resources.

"X-Amzn-Client-Token": A client token used to correlate requests and responses.

Main.Greengrass.create_software_update_job — Method

create_software_update_job(s3_url_signer_role, software_to_update, update_targets, update_targets_architecture, update_targets_operating_system)
create_software_update_job(s3_url_signer_role, software_to_update, update_targets, update_targets_architecture, update_targets_operating_system, params::Dict{String,<:Any})

Creates a software update for a core or group of cores (specified as an IoT thing group.) Use this to update the OTA Agent as well as the Greengrass core software. It makes use of the IoT Jobs feature which provides additional commands to manage a Greengrass core software update job.

Arguments

s3_url_signer_role:

software_to_update:

update_targets:

update_targets_architecture:

update_targets_operating_system:

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"UpdateAgentLogLevel":

"X-Amzn-Client-Token": A client token used to correlate requests and responses.

Main.Greengrass.create_subscription_definition — Method

create_subscription_definition()
create_subscription_definition(params::Dict{String,<:Any})

Creates a subscription definition. You may provide the initial version of the subscription definition now or use ''CreateSubscriptionDefinitionVersion'' at a later time.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"InitialVersion": Information about the initial version of the subscription definition.

"Name": The name of the subscription definition.

"X-Amzn-Client-Token": A client token used to correlate requests and responses.

"tags": Tag(s) to add to the new resource.

Main.Greengrass.create_subscription_definition_version — Method

create_subscription_definition_version(subscription_definition_id)
create_subscription_definition_version(subscription_definition_id, params::Dict{String,<:Any})

Creates a version of a subscription definition which has already been defined.

Arguments

subscription_definition_id: The ID of the subscription definition.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Subscriptions": A list of subscriptions.

"X-Amzn-Client-Token": A client token used to correlate requests and responses.

Main.Greengrass.delete_connector_definition — Method

delete_connector_definition(connector_definition_id)
delete_connector_definition(connector_definition_id, params::Dict{String,<:Any})

Deletes a connector definition.

Arguments

connector_definition_id: The ID of the connector definition.

Main.Greengrass.delete_core_definition — Method

delete_core_definition(core_definition_id)
delete_core_definition(core_definition_id, params::Dict{String,<:Any})

Deletes a core definition.

Arguments

core_definition_id: The ID of the core definition.

Main.Greengrass.delete_device_definition — Method

delete_device_definition(device_definition_id)
delete_device_definition(device_definition_id, params::Dict{String,<:Any})

Deletes a device definition.

Arguments

device_definition_id: The ID of the device definition.

Main.Greengrass.delete_function_definition — Method

delete_function_definition(function_definition_id)
delete_function_definition(function_definition_id, params::Dict{String,<:Any})

Deletes a Lambda function definition.

Arguments

function_definition_id: The ID of the Lambda function definition.

Main.Greengrass.delete_group — Method

delete_group(group_id)
delete_group(group_id, params::Dict{String,<:Any})

Deletes a group.

Arguments

group_id: The ID of the Greengrass group.

Main.Greengrass.delete_logger_definition — Method

delete_logger_definition(logger_definition_id)
delete_logger_definition(logger_definition_id, params::Dict{String,<:Any})

Deletes a logger definition.

Arguments

logger_definition_id: The ID of the logger definition.

Main.Greengrass.delete_resource_definition — Method

delete_resource_definition(resource_definition_id)
delete_resource_definition(resource_definition_id, params::Dict{String,<:Any})

Deletes a resource definition.

Arguments

resource_definition_id: The ID of the resource definition.

Main.Greengrass.delete_subscription_definition — Method

delete_subscription_definition(subscription_definition_id)
delete_subscription_definition(subscription_definition_id, params::Dict{String,<:Any})

Deletes a subscription definition.

Arguments

subscription_definition_id: The ID of the subscription definition.

Main.Greengrass.disassociate_role_from_group — Method

disassociate_role_from_group(group_id)
disassociate_role_from_group(group_id, params::Dict{String,<:Any})

Disassociates the role from a group.

Arguments

group_id: The ID of the Greengrass group.

Main.Greengrass.disassociate_service_role_from_account — Method

disassociate_service_role_from_account()
disassociate_service_role_from_account(params::Dict{String,<:Any})

Disassociates the service role from your account. Without a service role, deployments will not work.

Main.Greengrass.get_associated_role — Method

get_associated_role(group_id)
get_associated_role(group_id, params::Dict{String,<:Any})

Retrieves the role associated with a particular group.

Arguments

group_id: The ID of the Greengrass group.

Main.Greengrass.get_bulk_deployment_status — Method

get_bulk_deployment_status(bulk_deployment_id)
get_bulk_deployment_status(bulk_deployment_id, params::Dict{String,<:Any})

Returns the status of a bulk deployment.

Arguments

bulk_deployment_id: The ID of the bulk deployment.

Main.Greengrass.get_connectivity_info — Method

get_connectivity_info(thing_name)
get_connectivity_info(thing_name, params::Dict{String,<:Any})

Retrieves the connectivity information for a core.

Arguments

thing_name: The thing name.

Main.Greengrass.get_connector_definition — Method

get_connector_definition(connector_definition_id)
get_connector_definition(connector_definition_id, params::Dict{String,<:Any})

Retrieves information about a connector definition.

Arguments

connector_definition_id: The ID of the connector definition.

Main.Greengrass.get_connector_definition_version — Method

get_connector_definition_version(connector_definition_id, connector_definition_version_id)
get_connector_definition_version(connector_definition_id, connector_definition_version_id, params::Dict{String,<:Any})

Retrieves information about a connector definition version, including the connectors that the version contains. Connectors are prebuilt modules that interact with local infrastructure, device protocols, AWS, and other cloud services.

Arguments

connector_definition_id: The ID of the connector definition.

connector_definition_version_id: The ID of the connector definition version. This value maps to the ''Version'' property of the corresponding ''VersionInformation'' object, which is returned by ''ListConnectorDefinitionVersions'' requests. If the version is the last one that was associated with a connector definition, the value also maps to the ''LatestVersion'' property of the corresponding ''DefinitionInformation'' object.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"NextToken": The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.get_core_definition — Method

get_core_definition(core_definition_id)
get_core_definition(core_definition_id, params::Dict{String,<:Any})

Retrieves information about a core definition version.

Arguments

core_definition_id: The ID of the core definition.

Main.Greengrass.get_core_definition_version — Method

get_core_definition_version(core_definition_id, core_definition_version_id)
get_core_definition_version(core_definition_id, core_definition_version_id, params::Dict{String,<:Any})

Retrieves information about a core definition version.

Arguments

core_definition_id: The ID of the core definition.

core_definition_version_id: The ID of the core definition version. This value maps to the ''Version'' property of the corresponding ''VersionInformation'' object, which is returned by ''ListCoreDefinitionVersions'' requests. If the version is the last one that was associated with a core definition, the value also maps to the ''LatestVersion'' property of the corresponding ''DefinitionInformation'' object.

Main.Greengrass.get_deployment_status — Method

get_deployment_status(deployment_id, group_id)
get_deployment_status(deployment_id, group_id, params::Dict{String,<:Any})

Returns the status of a deployment.

Arguments

deployment_id: The ID of the deployment.

group_id: The ID of the Greengrass group.

Main.Greengrass.get_device_definition — Method

get_device_definition(device_definition_id)
get_device_definition(device_definition_id, params::Dict{String,<:Any})

Retrieves information about a device definition.

Arguments

device_definition_id: The ID of the device definition.

Main.Greengrass.get_device_definition_version — Method

get_device_definition_version(device_definition_id, device_definition_version_id)
get_device_definition_version(device_definition_id, device_definition_version_id, params::Dict{String,<:Any})

Retrieves information about a device definition version.

Arguments

device_definition_id: The ID of the device definition.

device_definition_version_id: The ID of the device definition version. This value maps to the ''Version'' property of the corresponding ''VersionInformation'' object, which is returned by ''ListDeviceDefinitionVersions'' requests. If the version is the last one that was associated with a device definition, the value also maps to the ''LatestVersion'' property of the corresponding ''DefinitionInformation'' object.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"NextToken": The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.get_function_definition — Method

get_function_definition(function_definition_id)
get_function_definition(function_definition_id, params::Dict{String,<:Any})

Retrieves information about a Lambda function definition, including its creation time and latest version.

Arguments

function_definition_id: The ID of the Lambda function definition.

Main.Greengrass.get_function_definition_version — Method

get_function_definition_version(function_definition_id, function_definition_version_id)
get_function_definition_version(function_definition_id, function_definition_version_id, params::Dict{String,<:Any})

Retrieves information about a Lambda function definition version, including which Lambda functions are included in the version and their configurations.

Arguments

function_definition_id: The ID of the Lambda function definition.

function_definition_version_id: The ID of the function definition version. This value maps to the ''Version'' property of the corresponding ''VersionInformation'' object, which is returned by ''ListFunctionDefinitionVersions'' requests. If the version is the last one that was associated with a function definition, the value also maps to the ''LatestVersion'' property of the corresponding ''DefinitionInformation'' object.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"NextToken": The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.get_group — Method

get_group(group_id)
get_group(group_id, params::Dict{String,<:Any})

Retrieves information about a group.

Arguments

group_id: The ID of the Greengrass group.

Main.Greengrass.get_group_certificate_authority — Method

get_group_certificate_authority(certificate_authority_id, group_id)
get_group_certificate_authority(certificate_authority_id, group_id, params::Dict{String,<:Any})

Retreives the CA associated with a group. Returns the public key of the CA.

Arguments

certificate_authority_id: The ID of the certificate authority.

group_id: The ID of the Greengrass group.

Main.Greengrass.get_group_certificate_configuration — Method

get_group_certificate_configuration(group_id)
get_group_certificate_configuration(group_id, params::Dict{String,<:Any})

Retrieves the current configuration for the CA used by the group.

Arguments

group_id: The ID of the Greengrass group.

Main.Greengrass.get_group_version — Method

get_group_version(group_id, group_version_id)
get_group_version(group_id, group_version_id, params::Dict{String,<:Any})

Retrieves information about a group version.

Arguments

group_id: The ID of the Greengrass group.

group_version_id: The ID of the group version. This value maps to the ''Version'' property of the corresponding ''VersionInformation'' object, which is returned by ''ListGroupVersions'' requests. If the version is the last one that was associated with a group, the value also maps to the ''LatestVersion'' property of the corresponding ''GroupInformation'' object.

Main.Greengrass.get_logger_definition — Method

get_logger_definition(logger_definition_id)
get_logger_definition(logger_definition_id, params::Dict{String,<:Any})

Retrieves information about a logger definition.

Arguments

logger_definition_id: The ID of the logger definition.

Main.Greengrass.get_logger_definition_version — Method

get_logger_definition_version(logger_definition_id, logger_definition_version_id)
get_logger_definition_version(logger_definition_id, logger_definition_version_id, params::Dict{String,<:Any})

Retrieves information about a logger definition version.

Arguments

logger_definition_id: The ID of the logger definition.

logger_definition_version_id: The ID of the logger definition version. This value maps to the ''Version'' property of the corresponding ''VersionInformation'' object, which is returned by ''ListLoggerDefinitionVersions'' requests. If the version is the last one that was associated with a logger definition, the value also maps to the ''LatestVersion'' property of the corresponding ''DefinitionInformation'' object.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"NextToken": The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.get_resource_definition — Method

get_resource_definition(resource_definition_id)
get_resource_definition(resource_definition_id, params::Dict{String,<:Any})

Retrieves information about a resource definition, including its creation time and latest version.

Arguments

resource_definition_id: The ID of the resource definition.

Main.Greengrass.get_resource_definition_version — Method

get_resource_definition_version(resource_definition_id, resource_definition_version_id)
get_resource_definition_version(resource_definition_id, resource_definition_version_id, params::Dict{String,<:Any})

Retrieves information about a resource definition version, including which resources are included in the version.

Arguments

resource_definition_id: The ID of the resource definition.

resource_definition_version_id: The ID of the resource definition version. This value maps to the ''Version'' property of the corresponding ''VersionInformation'' object, which is returned by ''ListResourceDefinitionVersions'' requests. If the version is the last one that was associated with a resource definition, the value also maps to the ''LatestVersion'' property of the corresponding ''DefinitionInformation'' object.

Main.Greengrass.get_service_role_for_account — Method

get_service_role_for_account()
get_service_role_for_account(params::Dict{String,<:Any})

Retrieves the service role that is attached to your account.

Main.Greengrass.get_subscription_definition — Method

get_subscription_definition(subscription_definition_id)
get_subscription_definition(subscription_definition_id, params::Dict{String,<:Any})

Retrieves information about a subscription definition.

Arguments

subscription_definition_id: The ID of the subscription definition.

Main.Greengrass.get_subscription_definition_version — Method

get_subscription_definition_version(subscription_definition_id, subscription_definition_version_id)
get_subscription_definition_version(subscription_definition_id, subscription_definition_version_id, params::Dict{String,<:Any})

Retrieves information about a subscription definition version.

Arguments

subscription_definition_id: The ID of the subscription definition.

subscription_definition_version_id: The ID of the subscription definition version. This value maps to the ''Version'' property of the corresponding ''VersionInformation'' object, which is returned by ''ListSubscriptionDefinitionVersions'' requests. If the version is the last one that was associated with a subscription definition, the value also maps to the ''LatestVersion'' property of the corresponding ''DefinitionInformation'' object.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"NextToken": The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.get_thing_runtime_configuration — Method

get_thing_runtime_configuration(thing_name)
get_thing_runtime_configuration(thing_name, params::Dict{String,<:Any})

Get the runtime configuration of a thing.

Arguments

thing_name: The thing name.

Main.Greengrass.list_bulk_deployment_detailed_reports — Method

list_bulk_deployment_detailed_reports(bulk_deployment_id)
list_bulk_deployment_detailed_reports(bulk_deployment_id, params::Dict{String,<:Any})

Gets a paginated list of the deployments that have been started in a bulk deployment operation, and their current deployment status.

Arguments

bulk_deployment_id: The ID of the bulk deployment.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to be returned per request.

"NextToken": The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_bulk_deployments — Method

list_bulk_deployments()
list_bulk_deployments(params::Dict{String,<:Any})

Returns a list of bulk deployments.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to be returned per request.

"NextToken": The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_connector_definition_versions — Method

list_connector_definition_versions(connector_definition_id)
list_connector_definition_versions(connector_definition_id, params::Dict{String,<:Any})

Lists the versions of a connector definition, which are containers for connectors. Connectors run on the Greengrass core and contain built-in integration with local infrastructure, device protocols, AWS, and other cloud services.

Arguments

connector_definition_id: The ID of the connector definition.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to be returned per request.

"NextToken": The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_connector_definitions — Method

list_connector_definitions()
list_connector_definitions(params::Dict{String,<:Any})

Retrieves a list of connector definitions.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to be returned per request.

"NextToken": The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_core_definition_versions — Method

list_core_definition_versions(core_definition_id)
list_core_definition_versions(core_definition_id, params::Dict{String,<:Any})

Lists the versions of a core definition.

Arguments

core_definition_id: The ID of the core definition.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to be returned per request.

"NextToken": The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_core_definitions — Method

list_core_definitions()
list_core_definitions(params::Dict{String,<:Any})

Retrieves a list of core definitions.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to be returned per request.

"NextToken": The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_deployments — Method

list_deployments(group_id)
list_deployments(group_id, params::Dict{String,<:Any})

Returns a history of deployments for the group.

Arguments

group_id: The ID of the Greengrass group.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to be returned per request.

"NextToken": The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_device_definition_versions — Method

list_device_definition_versions(device_definition_id)
list_device_definition_versions(device_definition_id, params::Dict{String,<:Any})

Lists the versions of a device definition.

Arguments

device_definition_id: The ID of the device definition.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to be returned per request.

"NextToken": The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_device_definitions — Method

list_device_definitions()
list_device_definitions(params::Dict{String,<:Any})

Retrieves a list of device definitions.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to be returned per request.

"NextToken": The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_function_definition_versions — Method

list_function_definition_versions(function_definition_id)
list_function_definition_versions(function_definition_id, params::Dict{String,<:Any})

Lists the versions of a Lambda function definition.

Arguments

function_definition_id: The ID of the Lambda function definition.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to be returned per request.

"NextToken": The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_function_definitions — Method

list_function_definitions()
list_function_definitions(params::Dict{String,<:Any})

Retrieves a list of Lambda function definitions.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to be returned per request.

"NextToken": The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_group_certificate_authorities — Method

list_group_certificate_authorities(group_id)
list_group_certificate_authorities(group_id, params::Dict{String,<:Any})

Retrieves the current CAs for a group.

Arguments

group_id: The ID of the Greengrass group.

Main.Greengrass.list_group_versions — Method

list_group_versions(group_id)
list_group_versions(group_id, params::Dict{String,<:Any})

Lists the versions of a group.

Arguments

group_id: The ID of the Greengrass group.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to be returned per request.

"NextToken": The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_groups — Method

list_groups()
list_groups(params::Dict{String,<:Any})

Retrieves a list of groups.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to be returned per request.

"NextToken": The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_logger_definition_versions — Method

list_logger_definition_versions(logger_definition_id)
list_logger_definition_versions(logger_definition_id, params::Dict{String,<:Any})

Lists the versions of a logger definition.

Arguments

logger_definition_id: The ID of the logger definition.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to be returned per request.

"NextToken": The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_logger_definitions — Method

list_logger_definitions()
list_logger_definitions(params::Dict{String,<:Any})

Retrieves a list of logger definitions.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to be returned per request.

"NextToken": The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_resource_definition_versions — Method

list_resource_definition_versions(resource_definition_id)
list_resource_definition_versions(resource_definition_id, params::Dict{String,<:Any})

Lists the versions of a resource definition.

Arguments

resource_definition_id: The ID of the resource definition.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to be returned per request.

"NextToken": The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_resource_definitions — Method

list_resource_definitions()
list_resource_definitions(params::Dict{String,<:Any})

Retrieves a list of resource definitions.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to be returned per request.

"NextToken": The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_subscription_definition_versions — Method

list_subscription_definition_versions(subscription_definition_id)
list_subscription_definition_versions(subscription_definition_id, params::Dict{String,<:Any})

Lists the versions of a subscription definition.

Arguments

subscription_definition_id: The ID of the subscription definition.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to be returned per request.

"NextToken": The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_subscription_definitions — Method

list_subscription_definitions()
list_subscription_definitions(params::Dict{String,<:Any})

Retrieves a list of subscription definitions.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to be returned per request.

"NextToken": The token for the next set of results, or ''null'' if there are no additional results.

Main.Greengrass.list_tags_for_resource — Method

list_tags_for_resource(resource-arn)
list_tags_for_resource(resource-arn, params::Dict{String,<:Any})

Retrieves a list of resource tags for a resource arn.

Arguments

resource-arn: The Amazon Resource Name (ARN) of the resource.

Main.Greengrass.reset_deployments — Method

reset_deployments(group_id)
reset_deployments(group_id, params::Dict{String,<:Any})

Resets a group's deployments.

Arguments

group_id: The ID of the Greengrass group.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Force": If true, performs a best-effort only core reset.

"X-Amzn-Client-Token": A client token used to correlate requests and responses.

Main.Greengrass.start_bulk_deployment — Method

start_bulk_deployment(execution_role_arn, input_file_uri)
start_bulk_deployment(execution_role_arn, input_file_uri, params::Dict{String,<:Any})

Deploys multiple groups in one operation. This action starts the bulk deployment of a specified set of group versions. Each group version deployment will be triggered with an adaptive rate that has a fixed upper limit. We recommend that you include an ''X-Amzn-Client-Token'' token in every ''StartBulkDeployment'' request. These requests are idempotent with respect to the token and the request parameters.

Arguments

execution_role_arn: The ARN of the execution role to associate with the bulk deployment operation. This IAM role must allow the ''greengrass:CreateDeployment'' action for all group versions that are listed in the input file. This IAM role must have access to the S3 bucket containing the input file.

input_file_uri: The URI of the input file contained in the S3 bucket. The execution role must have ''getObject'' permissions on this bucket to access the input file. The input file is a JSON-serialized, line delimited file with UTF-8 encoding that provides a list of group and version IDs and the deployment type. This file must be less than 100 MB. Currently, AWS IoT Greengrass supports only ''NewDeployment'' deployment types.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"X-Amzn-Client-Token": A client token used to correlate requests and responses.

"tags": Tag(s) to add to the new resource.

Main.Greengrass.stop_bulk_deployment — Method

stop_bulk_deployment(bulk_deployment_id)
stop_bulk_deployment(bulk_deployment_id, params::Dict{String,<:Any})

Stops the execution of a bulk deployment. This action returns a status of ''Stopping'' until the deployment is stopped. You cannot start a new bulk deployment while a previous deployment is in the ''Stopping'' state. This action doesn't rollback completed deployments or cancel pending deployments.

Arguments

bulk_deployment_id: The ID of the bulk deployment.

Main.Greengrass.tag_resource — Method

tag_resource(resource-arn)
tag_resource(resource-arn, params::Dict{String,<:Any})

Adds tags to a Greengrass resource. Valid resources are 'Group', 'ConnectorDefinition', 'CoreDefinition', 'DeviceDefinition', 'FunctionDefinition', 'LoggerDefinition', 'SubscriptionDefinition', 'ResourceDefinition', and 'BulkDeployment'.

Arguments

resource-arn: The Amazon Resource Name (ARN) of the resource.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"tags":

Main.Greengrass.untag_resource — Method

untag_resource(resource-arn, tag_keys)
untag_resource(resource-arn, tag_keys, params::Dict{String,<:Any})

Remove resource tags from a Greengrass Resource.

Arguments

resource-arn: The Amazon Resource Name (ARN) of the resource.

tag_keys: An array of tag keys to delete

Main.Greengrass.update_connectivity_info — Method

update_connectivity_info(thing_name)
update_connectivity_info(thing_name, params::Dict{String,<:Any})

Updates the connectivity information for the core. Any devices that belong to the group which has this core will receive this information in order to find the location of the core and connect to it.

Arguments

thing_name: The thing name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ConnectivityInfo": A list of connectivity info.

Main.Greengrass.update_connector_definition — Method

update_connector_definition(connector_definition_id)
update_connector_definition(connector_definition_id, params::Dict{String,<:Any})

Updates a connector definition.

Arguments

connector_definition_id: The ID of the connector definition.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Name": The name of the definition.

Main.Greengrass.update_core_definition — Method

update_core_definition(core_definition_id)
update_core_definition(core_definition_id, params::Dict{String,<:Any})

Updates a core definition.

Arguments

core_definition_id: The ID of the core definition.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Name": The name of the definition.

Main.Greengrass.update_device_definition — Method

update_device_definition(device_definition_id)
update_device_definition(device_definition_id, params::Dict{String,<:Any})

Updates a device definition.

Arguments

device_definition_id: The ID of the device definition.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Name": The name of the definition.

Main.Greengrass.update_function_definition — Method

update_function_definition(function_definition_id)
update_function_definition(function_definition_id, params::Dict{String,<:Any})

Updates a Lambda function definition.

Arguments

function_definition_id: The ID of the Lambda function definition.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Name": The name of the definition.

Main.Greengrass.update_group — Method

update_group(group_id)
update_group(group_id, params::Dict{String,<:Any})

Updates a group.

Arguments

group_id: The ID of the Greengrass group.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Name": The name of the definition.

Main.Greengrass.update_group_certificate_configuration — Method

update_group_certificate_configuration(group_id)
update_group_certificate_configuration(group_id, params::Dict{String,<:Any})

Updates the Certificate expiry time for a group.

Arguments

group_id: The ID of the Greengrass group.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"CertificateExpiryInMilliseconds": The amount of time remaining before the certificate expires, in milliseconds.

Main.Greengrass.update_logger_definition — Method

update_logger_definition(logger_definition_id)
update_logger_definition(logger_definition_id, params::Dict{String,<:Any})

Updates a logger definition.

Arguments

logger_definition_id: The ID of the logger definition.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Name": The name of the definition.

Main.Greengrass.update_resource_definition — Method

update_resource_definition(resource_definition_id)
update_resource_definition(resource_definition_id, params::Dict{String,<:Any})

Updates a resource definition.

Arguments

resource_definition_id: The ID of the resource definition.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Name": The name of the definition.

Main.Greengrass.update_subscription_definition — Method

update_subscription_definition(subscription_definition_id)
update_subscription_definition(subscription_definition_id, params::Dict{String,<:Any})

Updates a subscription definition.

Arguments

subscription_definition_id: The ID of the subscription definition.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Name": The name of the definition.

Main.Greengrass.update_thing_runtime_configuration — Method

update_thing_runtime_configuration(thing_name)
update_thing_runtime_configuration(thing_name, params::Dict{String,<:Any})

Updates the runtime configuration of a thing.

Arguments

thing_name: The thing name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"TelemetryConfiguration": Configuration for telemetry service.