Inspector2

This page documents function available when using the Inspector2 module, created with @service Inspector2.

Index

Main.Inspector2.associate_member

Main.Inspector2.batch_get_account_status

Main.Inspector2.batch_get_free_trial_info

Main.Inspector2.batch_get_member_ec2_deep_inspection_status

Main.Inspector2.batch_update_member_ec2_deep_inspection_status

Main.Inspector2.cancel_findings_report

Main.Inspector2.create_filter

Main.Inspector2.create_findings_report

Main.Inspector2.delete_filter

Main.Inspector2.describe_organization_configuration

Main.Inspector2.disable

Main.Inspector2.disable_delegated_admin_account

Main.Inspector2.disassociate_member

Main.Inspector2.enable

Main.Inspector2.enable_delegated_admin_account

Main.Inspector2.get_configuration

Main.Inspector2.get_delegated_admin_account

Main.Inspector2.get_ec2_deep_inspection_configuration

Main.Inspector2.get_findings_report_status

Main.Inspector2.get_member

Main.Inspector2.list_account_permissions

Main.Inspector2.list_coverage

Main.Inspector2.list_coverage_statistics

Main.Inspector2.list_delegated_admin_accounts

Main.Inspector2.list_filters

Main.Inspector2.list_finding_aggregations

Main.Inspector2.list_findings

Main.Inspector2.list_members

Main.Inspector2.list_tags_for_resource

Main.Inspector2.list_usage_totals

Main.Inspector2.search_vulnerabilities

Main.Inspector2.tag_resource

Main.Inspector2.untag_resource

Main.Inspector2.update_configuration

Main.Inspector2.update_ec2_deep_inspection_configuration

Main.Inspector2.update_filter

Main.Inspector2.update_org_ec2_deep_inspection_configuration

Main.Inspector2.update_organization_configuration

Documentation

Main.Inspector2.associate_member — Method

associate_member(account_id)
associate_member(account_id, params::Dict{String,<:Any})

Associates an Amazon Web Services account with an Amazon Inspector delegated administrator. An HTTP 200 response indicates the association was successfully started, but doesn’t indicate whether it was completed. You can check if the association completed by using ListMembers for multiple accounts or GetMembers for a single account.

Arguments

account_id: The Amazon Web Services account ID of the member account to be associated.

Main.Inspector2.batch_get_account_status — Method

batch_get_account_status()
batch_get_account_status(params::Dict{String,<:Any})

Retrieves the Amazon Inspector status of multiple Amazon Web Services accounts within your environment.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"accountIds": The 12-digit Amazon Web Services account IDs of the accounts to retrieve Amazon Inspector status for.

Main.Inspector2.batch_get_free_trial_info — Method

batch_get_free_trial_info(account_ids)
batch_get_free_trial_info(account_ids, params::Dict{String,<:Any})

Gets free trial status for multiple Amazon Web Services accounts.

Arguments

account_ids: The account IDs to get free trial status for.

Main.Inspector2.batch_get_member_ec2_deep_inspection_status — Method

batch_get_member_ec2_deep_inspection_status()
batch_get_member_ec2_deep_inspection_status(params::Dict{String,<:Any})

Retrieves Amazon Inspector deep inspection activation status of multiple member accounts within your organization. You must be the delegated administrator of an organization in Amazon Inspector to use this API.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"accountIds": The unique identifiers for the Amazon Web Services accounts to retrieve Amazon Inspector deep inspection activation status for. &lt;/p&gt;

Main.Inspector2.batch_update_member_ec2_deep_inspection_status — Method

batch_update_member_ec2_deep_inspection_status(account_ids)
batch_update_member_ec2_deep_inspection_status(account_ids, params::Dict{String,<:Any})

Activates or deactivates Amazon Inspector deep inspection for the provided member accounts in your organization. You must be the delegated administrator of an organization in Amazon Inspector to use this API.

Arguments

account_ids: The unique identifiers for the Amazon Web Services accounts to change Amazon Inspector deep inspection status for.

Main.Inspector2.cancel_findings_report — Method

cancel_findings_report(report_id)
cancel_findings_report(report_id, params::Dict{String,<:Any})

Cancels the given findings report.

Arguments

report_id: The ID of the report to be canceled.

Main.Inspector2.create_filter — Method

create_filter(action, filter_criteria, name)
create_filter(action, filter_criteria, name, params::Dict{String,<:Any})

Creates a filter resource using specified filter criteria.

Arguments

action: Defines the action that is to be applied to the findings that match the filter.

filter_criteria: Defines the criteria to be used in the filter for querying findings.

name: The name of the filter. Minimum length of 3. Maximum length of 64. Valid characters include alphanumeric characters, dot (.), underscore (_), and dash (-). Spaces are not allowed.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A description of the filter.

"reason": The reason for creating the filter.

"tags": A list of tags for the filter.

Main.Inspector2.create_findings_report — Method

create_findings_report(report_format, s3_destination)
create_findings_report(report_format, s3_destination, params::Dict{String,<:Any})

Creates a finding report. By default only ACTIVE findings are returned in the report. To see SUPRESSED or CLOSED findings you must specify a value for the findingStatus filter criteria.

Arguments

report_format: The format to generate the report in.

s3_destination: The Amazon S3 export destination for the report.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filterCriteria": The filter criteria to apply to the results of the finding report.

Main.Inspector2.delete_filter — Method

delete_filter(arn)
delete_filter(arn, params::Dict{String,<:Any})

Deletes a filter resource.

Arguments

arn: The Amazon Resource Number (ARN) of the filter to be deleted.

Main.Inspector2.describe_organization_configuration — Method

describe_organization_configuration()
describe_organization_configuration(params::Dict{String,<:Any})

Describe Amazon Inspector configuration settings for an Amazon Web Services organization.

Main.Inspector2.disable — Method

disable()
disable(params::Dict{String,<:Any})

Disables Amazon Inspector scans for one or more Amazon Web Services accounts. Disabling all scan types in an account disables the Amazon Inspector service.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"accountIds": An array of account IDs you want to disable Amazon Inspector scans for.

"resourceTypes": The resource scan types you want to disable.

Main.Inspector2.disable_delegated_admin_account — Method

disable_delegated_admin_account(delegated_admin_account_id)
disable_delegated_admin_account(delegated_admin_account_id, params::Dict{String,<:Any})

Disables the Amazon Inspector delegated administrator for your organization.

Arguments

delegated_admin_account_id: The Amazon Web Services account ID of the current Amazon Inspector delegated administrator.

Main.Inspector2.disassociate_member — Method

disassociate_member(account_id)
disassociate_member(account_id, params::Dict{String,<:Any})

Disassociates a member account from an Amazon Inspector delegated administrator.

Arguments

account_id: The Amazon Web Services account ID of the member account to disassociate.

Main.Inspector2.enable — Method

enable(resource_types)
enable(resource_types, params::Dict{String,<:Any})

Enables Amazon Inspector scans for one or more Amazon Web Services accounts.

Arguments

resource_types: The resource scan types you want to enable.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"accountIds": A list of account IDs you want to enable Amazon Inspector scans for.

"clientToken": The idempotency token for the request.

Main.Inspector2.enable_delegated_admin_account — Method

enable_delegated_admin_account(delegated_admin_account_id)
enable_delegated_admin_account(delegated_admin_account_id, params::Dict{String,<:Any})

Enables the Amazon Inspector delegated administrator for your Organizations organization.

Arguments

delegated_admin_account_id: The Amazon Web Services account ID of the Amazon Inspector delegated administrator.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": The idempotency token for the request.

Main.Inspector2.get_configuration — Method

get_configuration()
get_configuration(params::Dict{String,<:Any})

Retrieves setting configurations for Inspector scans.

Main.Inspector2.get_delegated_admin_account — Method

get_delegated_admin_account()
get_delegated_admin_account(params::Dict{String,<:Any})

Retrieves information about the Amazon Inspector delegated administrator for your organization.

Main.Inspector2.get_ec2_deep_inspection_configuration — Method

get_ec2_deep_inspection_configuration()
get_ec2_deep_inspection_configuration(params::Dict{String,<:Any})

Retrieves the activation status of Amazon Inspector deep inspection and custom paths associated with your account.

Main.Inspector2.get_findings_report_status — Method

get_findings_report_status()
get_findings_report_status(params::Dict{String,<:Any})

Gets the status of a findings report.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"reportId": The ID of the report to retrieve the status of.

Main.Inspector2.get_member — Method

get_member(account_id)
get_member(account_id, params::Dict{String,<:Any})

Gets member information for your organization.

Arguments

account_id: The Amazon Web Services account ID of the member account to retrieve information on.

Main.Inspector2.list_account_permissions — Method

list_account_permissions()
list_account_permissions(params::Dict{String,<:Any})

Lists the permissions an account has to configure Amazon Inspector.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return in the response.

"nextToken": A token to use for paginating results that are returned in the response. Set the value of this parameter to null for the first request to a list action. For subsequent calls, use the NextToken value returned from the previous request to continue listing results after the first page.

"service": The service scan type to check permissions for.

Main.Inspector2.list_coverage — Method

list_coverage()
list_coverage(params::Dict{String,<:Any})

Lists coverage details for you environment.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filterCriteria": An object that contains details on the filters to apply to the coverage data for your environment.

"maxResults": The maximum number of results to return in the response.

"nextToken": A token to use for paginating results that are returned in the response. Set the value of this parameter to null for the first request to a list action. For subsequent calls, use the NextToken value returned from the previous request to continue listing results after the first page.

Main.Inspector2.list_coverage_statistics — Method

list_coverage_statistics()
list_coverage_statistics(params::Dict{String,<:Any})

Lists Amazon Inspector coverage statistics for your environment.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filterCriteria": An object that contains details on the filters to apply to the coverage data for your environment.

"groupBy": The value to group the results by.

"nextToken": A token to use for paginating results that are returned in the response. Set the value of this parameter to null for the first request to a list action. For subsequent calls, use the NextToken value returned from the previous request to continue listing results after the first page.

Main.Inspector2.list_delegated_admin_accounts — Method

list_delegated_admin_accounts()
list_delegated_admin_accounts(params::Dict{String,<:Any})

Lists information about the Amazon Inspector delegated administrator of your organization.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return in the response.

"nextToken": A token to use for paginating results that are returned in the response. Set the value of this parameter to null for the first request to a list action. For subsequent calls, use the NextToken value returned from the previous request to continue listing results after the first page.

Main.Inspector2.list_filters — Method

list_filters()
list_filters(params::Dict{String,<:Any})

Lists the filters associated with your account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"action": The action the filter applies to matched findings.

"arns": The Amazon resource number (ARN) of the filter.

"maxResults": The maximum number of results to return in the response.

"nextToken": A token to use for paginating results that are returned in the response. Set the value of this parameter to null for the first request to a list action. For subsequent calls, use the NextToken value returned from the previous request to continue listing results after the first page.

Main.Inspector2.list_finding_aggregations — Method

list_finding_aggregations(aggregation_type)
list_finding_aggregations(aggregation_type, params::Dict{String,<:Any})

Lists aggregated finding data for your environment based on specific criteria.

Arguments

aggregation_type: The type of the aggregation request.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"accountIds": The Amazon Web Services account IDs to retrieve finding aggregation data for.

"aggregationRequest": Details of the aggregation request that is used to filter your aggregation results.

"maxResults": The maximum number of results to return in the response.

"nextToken": A token to use for paginating results that are returned in the response. Set the value of this parameter to null for the first request to a list action. For subsequent calls, use the NextToken value returned from the previous request to continue listing results after the first page.

Main.Inspector2.list_findings — Method

list_findings()
list_findings(params::Dict{String,<:Any})

Lists findings for your environment.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filterCriteria": Details on the filters to apply to your finding results.

"maxResults": The maximum number of results to return in the response.

"nextToken": A token to use for paginating results that are returned in the response. Set the value of this parameter to null for the first request to a list action. For subsequent calls, use the NextToken value returned from the previous request to continue listing results after the first page.

"sortCriteria": Details on the sort criteria to apply to your finding results.

Main.Inspector2.list_members — Method

list_members()
list_members(params::Dict{String,<:Any})

List members associated with the Amazon Inspector delegated administrator for your organization.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return in the response.

"nextToken": A token to use for paginating results that are returned in the response. Set the value of this parameter to null for the first request to a list action. For subsequent calls, use the NextToken value returned from the previous request to continue listing results after the first page.

"onlyAssociated": Specifies whether to list only currently associated members if True or to list all members within the organization if False.

Main.Inspector2.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists all tags attached to a given resource.

Arguments

resource_arn: The Amazon resource number (ARN) of the resource to list tags of.

Main.Inspector2.list_usage_totals — Method

list_usage_totals()
list_usage_totals(params::Dict{String,<:Any})

Lists the Amazon Inspector usage totals over the last 30 days.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"accountIds": The Amazon Web Services account IDs to retrieve usage totals for.

"maxResults": The maximum number of results to return in the response.

"nextToken": A token to use for paginating results that are returned in the response. Set the value of this parameter to null for the first request to a list action. For subsequent calls, use the NextToken value returned from the previous request to continue listing results after the first page.

Main.Inspector2.search_vulnerabilities — Method

search_vulnerabilities(filter_criteria)
search_vulnerabilities(filter_criteria, params::Dict{String,<:Any})

Lists Amazon Inspector coverage details for a specific vulnerability.

Arguments

filter_criteria: The criteria used to filter the results of a vulnerability search.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"nextToken": A token to use for paginating results that are returned in the response. Set the value of this parameter to null for the first request to a list action. For subsequent calls, use the NextToken value returned from the previous request to continue listing results after the first page.

Main.Inspector2.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Adds tags to a resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource to apply a tag to.

tags: The tags to be added to a resource.

Main.Inspector2.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes tags from a resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) for the resource to remove tags from.

tag_keys: The tag keys to remove from the resource.

Main.Inspector2.update_configuration — Method

update_configuration(ecr_configuration)
update_configuration(ecr_configuration, params::Dict{String,<:Any})

Updates setting configurations for your Amazon Inspector account. When you use this API as an Amazon Inspector delegated administrator this updates the setting for all accounts you manage. Member accounts in an organization cannot update this setting.

Arguments

ecr_configuration: Specifies how the ECR automated re-scan will be updated for your environment.

Main.Inspector2.update_ec2_deep_inspection_configuration — Method

update_ec2_deep_inspection_configuration()
update_ec2_deep_inspection_configuration(params::Dict{String,<:Any})

Activates, deactivates Amazon Inspector deep inspection, or updates custom paths for your account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"activateDeepInspection": Specify TRUE to activate Amazon Inspector deep inspection in your account, or FALSE to deactivate. Member accounts in an organization cannot deactivate deep inspection, instead the delegated administrator for the organization can deactivate a member account using BatchUpdateMemberEc2DeepInspectionStatus.

"packagePaths": The Amazon Inspector deep inspection custom paths you are adding for your account.

Main.Inspector2.update_filter — Method

update_filter(filter_arn)
update_filter(filter_arn, params::Dict{String,<:Any})

Specifies the action that is to be applied to the findings that match the filter.

Arguments

filter_arn: The Amazon Resource Number (ARN) of the filter to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"action": Specifies the action that is to be applied to the findings that match the filter.

"description": A description of the filter.

"filterCriteria": Defines the criteria to be update in the filter.

"name": The name of the filter.

"reason": The reason the filter was updated.

Main.Inspector2.update_org_ec2_deep_inspection_configuration — Method

update_org_ec2_deep_inspection_configuration(org_package_paths)
update_org_ec2_deep_inspection_configuration(org_package_paths, params::Dict{String,<:Any})

Updates the Amazon Inspector deep inspection custom paths for your organization. You must be an Amazon Inspector delegated administrator to use this API.

Arguments

org_package_paths: The Amazon Inspector deep inspection custom paths you are adding for your organization.

Main.Inspector2.update_organization_configuration — Method

update_organization_configuration(auto_enable)
update_organization_configuration(auto_enable, params::Dict{String,<:Any})

Updates the configurations for your Amazon Inspector organization.

Arguments

auto_enable: Defines which scan types are enabled automatically for new members of your Amazon Inspector organization.