Iot Wireless

This page documents function available when using the Iot_Wireless module, created with @service Iot_Wireless.

Index

Main.Iot_Wireless.associate_aws_account_with_partner_account

Main.Iot_Wireless.associate_multicast_group_with_fuota_task

Main.Iot_Wireless.associate_wireless_device_with_fuota_task

Main.Iot_Wireless.associate_wireless_device_with_multicast_group

Main.Iot_Wireless.associate_wireless_device_with_thing

Main.Iot_Wireless.associate_wireless_gateway_with_certificate

Main.Iot_Wireless.associate_wireless_gateway_with_thing

Main.Iot_Wireless.cancel_multicast_group_session

Main.Iot_Wireless.create_destination

Main.Iot_Wireless.create_device_profile

Main.Iot_Wireless.create_fuota_task

Main.Iot_Wireless.create_multicast_group

Main.Iot_Wireless.create_network_analyzer_configuration

Main.Iot_Wireless.create_service_profile

Main.Iot_Wireless.create_wireless_device

Main.Iot_Wireless.create_wireless_gateway

Main.Iot_Wireless.create_wireless_gateway_task

Main.Iot_Wireless.create_wireless_gateway_task_definition

Main.Iot_Wireless.delete_destination

Main.Iot_Wireless.delete_device_profile

Main.Iot_Wireless.delete_fuota_task

Main.Iot_Wireless.delete_multicast_group

Main.Iot_Wireless.delete_network_analyzer_configuration

Main.Iot_Wireless.delete_queued_messages

Main.Iot_Wireless.delete_service_profile

Main.Iot_Wireless.delete_wireless_device

Main.Iot_Wireless.delete_wireless_device_import_task

Main.Iot_Wireless.delete_wireless_gateway

Main.Iot_Wireless.delete_wireless_gateway_task

Main.Iot_Wireless.delete_wireless_gateway_task_definition

Main.Iot_Wireless.deregister_wireless_device

Main.Iot_Wireless.disassociate_aws_account_from_partner_account

Main.Iot_Wireless.disassociate_multicast_group_from_fuota_task

Main.Iot_Wireless.disassociate_wireless_device_from_fuota_task

Main.Iot_Wireless.disassociate_wireless_device_from_multicast_group

Main.Iot_Wireless.disassociate_wireless_device_from_thing

Main.Iot_Wireless.disassociate_wireless_gateway_from_certificate

Main.Iot_Wireless.disassociate_wireless_gateway_from_thing

Main.Iot_Wireless.get_destination

Main.Iot_Wireless.get_device_profile

Main.Iot_Wireless.get_event_configuration_by_resource_types

Main.Iot_Wireless.get_fuota_task

Main.Iot_Wireless.get_log_levels_by_resource_types

Main.Iot_Wireless.get_multicast_group

Main.Iot_Wireless.get_multicast_group_session

Main.Iot_Wireless.get_network_analyzer_configuration

Main.Iot_Wireless.get_partner_account

Main.Iot_Wireless.get_position

Main.Iot_Wireless.get_position_configuration

Main.Iot_Wireless.get_position_estimate

Main.Iot_Wireless.get_resource_event_configuration

Main.Iot_Wireless.get_resource_log_level

Main.Iot_Wireless.get_resource_position

Main.Iot_Wireless.get_service_endpoint

Main.Iot_Wireless.get_service_profile

Main.Iot_Wireless.get_wireless_device

Main.Iot_Wireless.get_wireless_device_import_task

Main.Iot_Wireless.get_wireless_device_statistics

Main.Iot_Wireless.get_wireless_gateway

Main.Iot_Wireless.get_wireless_gateway_certificate

Main.Iot_Wireless.get_wireless_gateway_firmware_information

Main.Iot_Wireless.get_wireless_gateway_statistics

Main.Iot_Wireless.get_wireless_gateway_task

Main.Iot_Wireless.get_wireless_gateway_task_definition

Main.Iot_Wireless.list_destinations

Main.Iot_Wireless.list_device_profiles

Main.Iot_Wireless.list_devices_for_wireless_device_import_task

Main.Iot_Wireless.list_event_configurations

Main.Iot_Wireless.list_fuota_tasks

Main.Iot_Wireless.list_multicast_groups

Main.Iot_Wireless.list_multicast_groups_by_fuota_task

Main.Iot_Wireless.list_network_analyzer_configurations

Main.Iot_Wireless.list_partner_accounts

Main.Iot_Wireless.list_position_configurations

Main.Iot_Wireless.list_queued_messages

Main.Iot_Wireless.list_service_profiles

Main.Iot_Wireless.list_tags_for_resource

Main.Iot_Wireless.list_wireless_device_import_tasks

Main.Iot_Wireless.list_wireless_devices

Main.Iot_Wireless.list_wireless_gateway_task_definitions

Main.Iot_Wireless.list_wireless_gateways

Main.Iot_Wireless.put_position_configuration

Main.Iot_Wireless.put_resource_log_level

Main.Iot_Wireless.reset_all_resource_log_levels

Main.Iot_Wireless.reset_resource_log_level

Main.Iot_Wireless.send_data_to_multicast_group

Main.Iot_Wireless.send_data_to_wireless_device

Main.Iot_Wireless.start_bulk_associate_wireless_device_with_multicast_group

Main.Iot_Wireless.start_bulk_disassociate_wireless_device_from_multicast_group

Main.Iot_Wireless.start_fuota_task

Main.Iot_Wireless.start_multicast_group_session

Main.Iot_Wireless.start_single_wireless_device_import_task

Main.Iot_Wireless.start_wireless_device_import_task

Main.Iot_Wireless.tag_resource

Main.Iot_Wireless.test_wireless_device

Main.Iot_Wireless.untag_resource

Main.Iot_Wireless.update_destination

Main.Iot_Wireless.update_event_configuration_by_resource_types

Main.Iot_Wireless.update_fuota_task

Main.Iot_Wireless.update_log_levels_by_resource_types

Main.Iot_Wireless.update_multicast_group

Main.Iot_Wireless.update_network_analyzer_configuration

Main.Iot_Wireless.update_partner_account

Main.Iot_Wireless.update_position

Main.Iot_Wireless.update_resource_event_configuration

Main.Iot_Wireless.update_resource_position

Main.Iot_Wireless.update_wireless_device

Main.Iot_Wireless.update_wireless_device_import_task

Main.Iot_Wireless.update_wireless_gateway

Documentation

Main.Iot_Wireless.associate_aws_account_with_partner_account — Method

associate_aws_account_with_partner_account(sidewalk)
associate_aws_account_with_partner_account(sidewalk, params::Dict{String,<:Any})

Associates a partner account with your AWS account.

Arguments

sidewalk: The Sidewalk account credentials.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientRequestToken": Each resource must have a unique client request token. If you try to create a new resource with the same token as a resource that already exists, an exception occurs. If you omit this value, AWS SDKs will automatically generate a unique client request.

"Tags": The tags to attach to the specified resource. Tags are metadata that you can use to manage a resource.

Main.Iot_Wireless.associate_multicast_group_with_fuota_task — Method

associate_multicast_group_with_fuota_task(id, multicast_group_id)
associate_multicast_group_with_fuota_task(id, multicast_group_id, params::Dict{String,<:Any})

Associate a multicast group with a FUOTA task.

Arguments

id:

multicast_group_id:

Main.Iot_Wireless.associate_wireless_device_with_fuota_task — Method

associate_wireless_device_with_fuota_task(id, wireless_device_id)
associate_wireless_device_with_fuota_task(id, wireless_device_id, params::Dict{String,<:Any})

Associate a wireless device with a FUOTA task.

Arguments

id:

wireless_device_id:

Main.Iot_Wireless.associate_wireless_device_with_multicast_group — Method

associate_wireless_device_with_multicast_group(id, wireless_device_id)
associate_wireless_device_with_multicast_group(id, wireless_device_id, params::Dict{String,<:Any})

Associates a wireless device with a multicast group.

Arguments

id:

wireless_device_id:

Main.Iot_Wireless.associate_wireless_device_with_thing — Method

associate_wireless_device_with_thing(id, thing_arn)
associate_wireless_device_with_thing(id, thing_arn, params::Dict{String,<:Any})

Associates a wireless device with a thing.

Arguments

id: The ID of the resource to update.

thing_arn: The ARN of the thing to associate with the wireless device.

Main.Iot_Wireless.associate_wireless_gateway_with_certificate — Method

associate_wireless_gateway_with_certificate(id, iot_certificate_id)
associate_wireless_gateway_with_certificate(id, iot_certificate_id, params::Dict{String,<:Any})

Associates a wireless gateway with a certificate.

Arguments

id: The ID of the resource to update.

iot_certificate_id: The ID of the certificate to associate with the wireless gateway.

Main.Iot_Wireless.associate_wireless_gateway_with_thing — Method

associate_wireless_gateway_with_thing(id, thing_arn)
associate_wireless_gateway_with_thing(id, thing_arn, params::Dict{String,<:Any})

Associates a wireless gateway with a thing.

Arguments

id: The ID of the resource to update.

thing_arn: The ARN of the thing to associate with the wireless gateway.

Main.Iot_Wireless.cancel_multicast_group_session — Method

cancel_multicast_group_session(id)
cancel_multicast_group_session(id, params::Dict{String,<:Any})

Cancels an existing multicast group session.

Arguments

id:

Main.Iot_Wireless.create_destination — Method

create_destination(expression, expression_type, name, role_arn)
create_destination(expression, expression_type, name, role_arn, params::Dict{String,<:Any})

Creates a new destination that maps a device message to an AWS IoT rule.

Arguments

expression: The rule name or topic rule to send messages to.

expression_type: The type of value in Expression.

name: The name of the new resource.

role_arn: The ARN of the IAM Role that authorizes the destination.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientRequestToken": Each resource must have a unique client request token. If you try to create a new resource with the same token as a resource that already exists, an exception occurs. If you omit this value, AWS SDKs will automatically generate a unique client request.

"Description": The description of the new resource.

"Tags": The tags to attach to the new destination. Tags are metadata that you can use to manage a resource.

Main.Iot_Wireless.create_device_profile — Method

create_device_profile()
create_device_profile(params::Dict{String,<:Any})

Creates a new device profile.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientRequestToken": Each resource must have a unique client request token. If you try to create a new resource with the same token as a resource that already exists, an exception occurs. If you omit this value, AWS SDKs will automatically generate a unique client request.

"LoRaWAN": The device profile information to use to create the device profile.

"Name": The name of the new resource.

"Sidewalk": The Sidewalk-related information for creating the Sidewalk device profile.

"Tags": The tags to attach to the new device profile. Tags are metadata that you can use to manage a resource.

Main.Iot_Wireless.create_fuota_task — Method

create_fuota_task(firmware_update_image, firmware_update_role)
create_fuota_task(firmware_update_image, firmware_update_role, params::Dict{String,<:Any})

Creates a FUOTA task.

Arguments

firmware_update_image:

firmware_update_role:

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientRequestToken":

"Description":

"FragmentIntervalMS":

"FragmentSizeBytes":

"LoRaWAN":

"Name":

"RedundancyPercent":

"Tags":

Main.Iot_Wireless.create_multicast_group — Method

create_multicast_group(lo_ra_wan)
create_multicast_group(lo_ra_wan, params::Dict{String,<:Any})

Creates a multicast group.

Arguments

lo_ra_wan:

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientRequestToken": Each resource must have a unique client request token. If you try to create a new resource with the same token as a resource that already exists, an exception occurs. If you omit this value, AWS SDKs will automatically generate a unique client request.

"Description": The description of the multicast group.

"Name":

"Tags":

Main.Iot_Wireless.create_network_analyzer_configuration — Method

create_network_analyzer_configuration(name)
create_network_analyzer_configuration(name, params::Dict{String,<:Any})

Creates a new network analyzer configuration.

Arguments

name:

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientRequestToken":

"Description":

"MulticastGroups": Multicast Group resources to add to the network analyzer configruation. Provide the MulticastGroupId of the resource to add in the input array.

"Tags":

"TraceContent":

"WirelessDevices": Wireless device resources to add to the network analyzer configuration. Provide the WirelessDeviceId of the resource to add in the input array.

"WirelessGateways": Wireless gateway resources to add to the network analyzer configuration. Provide the WirelessGatewayId of the resource to add in the input array.

Main.Iot_Wireless.create_service_profile — Method

create_service_profile()
create_service_profile(params::Dict{String,<:Any})

Creates a new service profile.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientRequestToken": Each resource must have a unique client request token. If you try to create a new resource with the same token as a resource that already exists, an exception occurs. If you omit this value, AWS SDKs will automatically generate a unique client request.

"LoRaWAN": The service profile information to use to create the service profile.

"Name": The name of the new resource.

"Tags": The tags to attach to the new service profile. Tags are metadata that you can use to manage a resource.

Main.Iot_Wireless.create_wireless_device — Method

create_wireless_device(destination_name, type)
create_wireless_device(destination_name, type, params::Dict{String,<:Any})

Provisions a wireless device.

Arguments

destination_name: The name of the destination to assign to the new wireless device.

type: The wireless device type.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientRequestToken": Each resource must have a unique client request token. If you try to create a new resource with the same token as a resource that already exists, an exception occurs. If you omit this value, AWS SDKs will automatically generate a unique client request.

"Description": The description of the new resource.

"LoRaWAN": The device configuration information to use to create the wireless device.

"Name": The name of the new resource.

"Positioning": FPort values for the GNSS, stream, and ClockSync functions of the positioning information.

"Sidewalk": The device configuration information to use to create the Sidewalk device.

"Tags": The tags to attach to the new wireless device. Tags are metadata that you can use to manage a resource.

Main.Iot_Wireless.create_wireless_gateway — Method

create_wireless_gateway(lo_ra_wan)
create_wireless_gateway(lo_ra_wan, params::Dict{String,<:Any})

Provisions a wireless gateway.

Arguments

lo_ra_wan: The gateway configuration information to use to create the wireless gateway.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientRequestToken": Each resource must have a unique client request token. If you try to create a new resource with the same token as a resource that already exists, an exception occurs. If you omit this value, AWS SDKs will automatically generate a unique client request.

"Description": The description of the new resource.

"Name": The name of the new resource.

"Tags": The tags to attach to the new wireless gateway. Tags are metadata that you can use to manage a resource.

Main.Iot_Wireless.create_wireless_gateway_task — Method

create_wireless_gateway_task(id, wireless_gateway_task_definition_id)
create_wireless_gateway_task(id, wireless_gateway_task_definition_id, params::Dict{String,<:Any})

Creates a task for a wireless gateway.

Arguments

id: The ID of the resource to update.

wireless_gateway_task_definition_id: The ID of the WirelessGatewayTaskDefinition.

Main.Iot_Wireless.create_wireless_gateway_task_definition — Method

create_wireless_gateway_task_definition(auto_create_tasks)
create_wireless_gateway_task_definition(auto_create_tasks, params::Dict{String,<:Any})

Creates a gateway task definition.

Arguments

auto_create_tasks: Whether to automatically create tasks using this task definition for all gateways with the specified current version. If false, the task must me created by calling CreateWirelessGatewayTask.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientRequestToken": Each resource must have a unique client request token. If you try to create a new resource with the same token as a resource that already exists, an exception occurs. If you omit this value, AWS SDKs will automatically generate a unique client request.

"Name": The name of the new resource.

"Tags": The tags to attach to the specified resource. Tags are metadata that you can use to manage a resource.

"Update": Information about the gateways to update.

Main.Iot_Wireless.delete_destination — Method

delete_destination(name)
delete_destination(name, params::Dict{String,<:Any})

Deletes a destination.

Arguments

name: The name of the resource to delete.

Main.Iot_Wireless.delete_device_profile — Method

delete_device_profile(id)
delete_device_profile(id, params::Dict{String,<:Any})

Deletes a device profile.

Arguments

id: The ID of the resource to delete.

Main.Iot_Wireless.delete_fuota_task — Method

delete_fuota_task(id)
delete_fuota_task(id, params::Dict{String,<:Any})

Deletes a FUOTA task.

Arguments

id:

Main.Iot_Wireless.delete_multicast_group — Method

delete_multicast_group(id)
delete_multicast_group(id, params::Dict{String,<:Any})

Deletes a multicast group if it is not in use by a fuota task.

Arguments

id:

Main.Iot_Wireless.delete_network_analyzer_configuration — Method

delete_network_analyzer_configuration(configuration_name)
delete_network_analyzer_configuration(configuration_name, params::Dict{String,<:Any})

Deletes a network analyzer configuration.

Arguments

configuration_name:

Main.Iot_Wireless.delete_queued_messages — Method

delete_queued_messages(id, message_id)
delete_queued_messages(id, message_id, params::Dict{String,<:Any})

Remove queued messages from the downlink queue.

Arguments

id: The ID of a given wireless device for which downlink messages will be deleted.

message_id: If message ID is "*", it cleares the entire downlink queue for a given device, specified by the wireless device ID. Otherwise, the downlink message with the specified message ID will be deleted.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"WirelessDeviceType": The wireless device type, which can be either Sidewalk or LoRaWAN.

Main.Iot_Wireless.delete_service_profile — Method

delete_service_profile(id)
delete_service_profile(id, params::Dict{String,<:Any})

Deletes a service profile.

Arguments

id: The ID of the resource to delete.

Main.Iot_Wireless.delete_wireless_device — Method

delete_wireless_device(id)
delete_wireless_device(id, params::Dict{String,<:Any})

Deletes a wireless device.

Arguments

id: The ID of the resource to delete.

Main.Iot_Wireless.delete_wireless_device_import_task — Method

delete_wireless_device_import_task(id)
delete_wireless_device_import_task(id, params::Dict{String,<:Any})

Delete an import task.

Arguments

id: The unique identifier of the import task to be deleted.

Main.Iot_Wireless.delete_wireless_gateway — Method

delete_wireless_gateway(id)
delete_wireless_gateway(id, params::Dict{String,<:Any})

Deletes a wireless gateway.

Arguments

id: The ID of the resource to delete.

Main.Iot_Wireless.delete_wireless_gateway_task — Method

delete_wireless_gateway_task(id)
delete_wireless_gateway_task(id, params::Dict{String,<:Any})

Deletes a wireless gateway task.

Arguments

id: The ID of the resource to delete.

Main.Iot_Wireless.delete_wireless_gateway_task_definition — Method

delete_wireless_gateway_task_definition(id)
delete_wireless_gateway_task_definition(id, params::Dict{String,<:Any})

Deletes a wireless gateway task definition. Deleting this task definition does not affect tasks that are currently in progress.

Arguments

id: The ID of the resource to delete.

Main.Iot_Wireless.deregister_wireless_device — Method

deregister_wireless_device(identifier)
deregister_wireless_device(identifier, params::Dict{String,<:Any})

Deregister a wireless device from AWS IoT Wireless.

Arguments

identifier: The identifier of the wireless device to deregister from AWS IoT Wireless.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"WirelessDeviceType": The type of wireless device to deregister from AWS IoT Wireless, which can be LoRaWAN or Sidewalk.

Main.Iot_Wireless.disassociate_aws_account_from_partner_account — Method

disassociate_aws_account_from_partner_account(partner_account_id, partner_type)
disassociate_aws_account_from_partner_account(partner_account_id, partner_type, params::Dict{String,<:Any})

Disassociates your AWS account from a partner account. If PartnerAccountId and PartnerType are null, disassociates your AWS account from all partner accounts.

Arguments

partner_account_id: The partner account ID to disassociate from the AWS account.

partner_type: The partner type.

Main.Iot_Wireless.disassociate_multicast_group_from_fuota_task — Method

disassociate_multicast_group_from_fuota_task(id, multicast_group_id)
disassociate_multicast_group_from_fuota_task(id, multicast_group_id, params::Dict{String,<:Any})

Disassociates a multicast group from a fuota task.

Arguments

id:

multicast_group_id:

Main.Iot_Wireless.disassociate_wireless_device_from_fuota_task — Method

disassociate_wireless_device_from_fuota_task(id, wireless_device_id)
disassociate_wireless_device_from_fuota_task(id, wireless_device_id, params::Dict{String,<:Any})

Disassociates a wireless device from a FUOTA task.

Arguments

id:

wireless_device_id:

Main.Iot_Wireless.disassociate_wireless_device_from_multicast_group — Method

disassociate_wireless_device_from_multicast_group(id, wireless_device_id)
disassociate_wireless_device_from_multicast_group(id, wireless_device_id, params::Dict{String,<:Any})

Disassociates a wireless device from a multicast group.

Arguments

id:

wireless_device_id:

Main.Iot_Wireless.disassociate_wireless_device_from_thing — Method

disassociate_wireless_device_from_thing(id)
disassociate_wireless_device_from_thing(id, params::Dict{String,<:Any})

Disassociates a wireless device from its currently associated thing.

Arguments

id: The ID of the resource to update.

Main.Iot_Wireless.disassociate_wireless_gateway_from_certificate — Method

disassociate_wireless_gateway_from_certificate(id)
disassociate_wireless_gateway_from_certificate(id, params::Dict{String,<:Any})

Disassociates a wireless gateway from its currently associated certificate.

Arguments

id: The ID of the resource to update.

Main.Iot_Wireless.disassociate_wireless_gateway_from_thing — Method

disassociate_wireless_gateway_from_thing(id)
disassociate_wireless_gateway_from_thing(id, params::Dict{String,<:Any})

Disassociates a wireless gateway from its currently associated thing.

Arguments

id: The ID of the resource to update.

Main.Iot_Wireless.get_destination — Method

get_destination(name)
get_destination(name, params::Dict{String,<:Any})

Gets information about a destination.

Arguments

name: The name of the resource to get.

Main.Iot_Wireless.get_device_profile — Method

get_device_profile(id)
get_device_profile(id, params::Dict{String,<:Any})

Gets information about a device profile.

Arguments

id: The ID of the resource to get.

Main.Iot_Wireless.get_event_configuration_by_resource_types — Method

get_event_configuration_by_resource_types()
get_event_configuration_by_resource_types(params::Dict{String,<:Any})

Get the event configuration based on resource types.

Main.Iot_Wireless.get_fuota_task — Method

get_fuota_task(id)
get_fuota_task(id, params::Dict{String,<:Any})

Gets information about a FUOTA task.

Arguments

id:

Main.Iot_Wireless.get_log_levels_by_resource_types — Method

get_log_levels_by_resource_types()
get_log_levels_by_resource_types(params::Dict{String,<:Any})

Returns current default log levels or log levels by resource types. Based on resource types, log levels can be for wireless device log options or wireless gateway log options.

Main.Iot_Wireless.get_multicast_group — Method

get_multicast_group(id)
get_multicast_group(id, params::Dict{String,<:Any})

Gets information about a multicast group.

Arguments

id:

Main.Iot_Wireless.get_multicast_group_session — Method

get_multicast_group_session(id)
get_multicast_group_session(id, params::Dict{String,<:Any})

Gets information about a multicast group session.

Arguments

id:

Main.Iot_Wireless.get_network_analyzer_configuration — Method

get_network_analyzer_configuration(configuration_name)
get_network_analyzer_configuration(configuration_name, params::Dict{String,<:Any})

Get network analyzer configuration.

Arguments

configuration_name:

Main.Iot_Wireless.get_partner_account — Method

get_partner_account(partner_account_id, partner_type)
get_partner_account(partner_account_id, partner_type, params::Dict{String,<:Any})

Gets information about a partner account. If PartnerAccountId and PartnerType are null, returns all partner accounts.

Arguments

partner_account_id: The partner account ID to disassociate from the AWS account.

partner_type: The partner type.

Main.Iot_Wireless.get_position — Method

get_position(resource_identifier, resource_type)
get_position(resource_identifier, resource_type, params::Dict{String,<:Any})

Get the position information for a given resource. This action is no longer supported. Calls to retrieve the position information should use the GetResourcePosition API operation instead.

Arguments

resource_identifier: Resource identifier used to retrieve the position information.

resource_type: Resource type of the resource for which position information is retrieved.

Main.Iot_Wireless.get_position_configuration — Method

get_position_configuration(resource_identifier, resource_type)
get_position_configuration(resource_identifier, resource_type, params::Dict{String,<:Any})

Get position configuration for a given resource. This action is no longer supported. Calls to retrieve the position configuration should use the GetResourcePosition API operation instead.

Arguments

resource_identifier: Resource identifier used in a position configuration.

resource_type: Resource type of the resource for which position configuration is retrieved.

Main.Iot_Wireless.get_position_estimate — Method

get_position_estimate()
get_position_estimate(params::Dict{String,<:Any})

Get estimated position information as a payload in GeoJSON format. The payload measurement data is resolved using solvers that are provided by third-party vendors.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"CellTowers": Retrieves an estimated device position by resolving measurement data from cellular radio towers. The position is resolved using HERE's cellular-based solver.

"Gnss": Retrieves an estimated device position by resolving the global navigation satellite system (GNSS) scan data. The position is resolved using the GNSS solver powered by LoRa Cloud.

"Ip": Retrieves an estimated device position by resolving the IP address information from the device. The position is resolved using MaxMind's IP-based solver.

"Timestamp": Optional information that specifies the time when the position information will be resolved. It uses the Unix timestamp format. If not specified, the time at which the request was received will be used.

"WiFiAccessPoints": Retrieves an estimated device position by resolving WLAN measurement data. The position is resolved using HERE's Wi-Fi based solver.

Main.Iot_Wireless.get_resource_event_configuration — Method

get_resource_event_configuration(identifier, identifier_type)
get_resource_event_configuration(identifier, identifier_type, params::Dict{String,<:Any})

Get the event configuration for a particular resource identifier.

Arguments

identifier: Resource identifier to opt in for event messaging.

identifier_type: Identifier type of the particular resource identifier for event configuration.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"partnerType": Partner type of the resource if the identifier type is PartnerAccountId.

Main.Iot_Wireless.get_resource_log_level — Method

get_resource_log_level(resource_identifier, resource_type)
get_resource_log_level(resource_identifier, resource_type, params::Dict{String,<:Any})

Fetches the log-level override, if any, for a given resource-ID and resource-type. It can be used for a wireless device or a wireless gateway.

Arguments

resource_identifier:

resource_type: The type of the resource, which can be WirelessDevice or WirelessGateway.

Main.Iot_Wireless.get_resource_position — Method

get_resource_position(resource_identifier, resource_type)
get_resource_position(resource_identifier, resource_type, params::Dict{String,<:Any})

Get the position information for a given wireless device or a wireless gateway resource. The position information uses the World Geodetic System (WGS84).

Arguments

resource_identifier: The identifier of the resource for which position information is retrieved. It can be the wireless device ID or the wireless gateway ID, depending on the resource type.

resource_type: The type of resource for which position information is retrieved, which can be a wireless device or a wireless gateway.

Main.Iot_Wireless.get_service_endpoint — Method

get_service_endpoint()
get_service_endpoint(params::Dict{String,<:Any})

Gets the account-specific endpoint for Configuration and Update Server (CUPS) protocol or LoRaWAN Network Server (LNS) connections.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"serviceType": The service type for which to get endpoint information about. Can be CUPS for the Configuration and Update Server endpoint, or LNS for the LoRaWAN Network Server endpoint or CLAIM for the global endpoint.

Main.Iot_Wireless.get_service_profile — Method

get_service_profile(id)
get_service_profile(id, params::Dict{String,<:Any})

Gets information about a service profile.

Arguments

id: The ID of the resource to get.

Main.Iot_Wireless.get_wireless_device — Method

get_wireless_device(identifier, identifier_type)
get_wireless_device(identifier, identifier_type, params::Dict{String,<:Any})

Gets information about a wireless device.

Arguments

identifier: The identifier of the wireless device to get.

identifier_type: The type of identifier used in identifier.

Main.Iot_Wireless.get_wireless_device_import_task — Method

get_wireless_device_import_task(id)
get_wireless_device_import_task(id, params::Dict{String,<:Any})

Get information about an import task and count of device onboarding summary information for the import task.

Arguments

id: The identifier of the import task for which information is requested.

Main.Iot_Wireless.get_wireless_device_statistics — Method

get_wireless_device_statistics(id)
get_wireless_device_statistics(id, params::Dict{String,<:Any})

Gets operating information about a wireless device.

Arguments

id: The ID of the wireless device for which to get the data.

Main.Iot_Wireless.get_wireless_gateway — Method

get_wireless_gateway(identifier, identifier_type)
get_wireless_gateway(identifier, identifier_type, params::Dict{String,<:Any})

Gets information about a wireless gateway.

Arguments

identifier: The identifier of the wireless gateway to get.

identifier_type: The type of identifier used in identifier.

Main.Iot_Wireless.get_wireless_gateway_certificate — Method

get_wireless_gateway_certificate(id)
get_wireless_gateway_certificate(id, params::Dict{String,<:Any})

Gets the ID of the certificate that is currently associated with a wireless gateway.

Arguments

id: The ID of the resource to get.

Main.Iot_Wireless.get_wireless_gateway_firmware_information — Method

get_wireless_gateway_firmware_information(id)
get_wireless_gateway_firmware_information(id, params::Dict{String,<:Any})

Gets the firmware version and other information about a wireless gateway.

Arguments

id: The ID of the resource to get.

Main.Iot_Wireless.get_wireless_gateway_statistics — Method

get_wireless_gateway_statistics(id)
get_wireless_gateway_statistics(id, params::Dict{String,<:Any})

Gets operating information about a wireless gateway.

Arguments

id: The ID of the wireless gateway for which to get the data.

Main.Iot_Wireless.get_wireless_gateway_task — Method

get_wireless_gateway_task(id)
get_wireless_gateway_task(id, params::Dict{String,<:Any})

Gets information about a wireless gateway task.

Arguments

id: The ID of the resource to get.

Main.Iot_Wireless.get_wireless_gateway_task_definition — Method

get_wireless_gateway_task_definition(id)
get_wireless_gateway_task_definition(id, params::Dict{String,<:Any})

Gets information about a wireless gateway task definition.

Arguments

id: The ID of the resource to get.

Main.Iot_Wireless.list_destinations — Method

list_destinations()
list_destinations(params::Dict{String,<:Any})

Lists the destinations registered to your AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return in this operation.

"nextToken": To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

Main.Iot_Wireless.list_device_profiles — Method

list_device_profiles()
list_device_profiles(params::Dict{String,<:Any})

Lists the device profiles registered to your AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"deviceProfileType": A filter to list only device profiles that use this type, which can be LoRaWAN or Sidewalk.

"maxResults": The maximum number of results to return in this operation.

"nextToken": To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

Main.Iot_Wireless.list_devices_for_wireless_device_import_task — Method

list_devices_for_wireless_device_import_task(id)
list_devices_for_wireless_device_import_task(id, params::Dict{String,<:Any})

List the Sidewalk devices in an import task and their onboarding status.

Arguments

id: The identifier of the import task for which wireless devices are listed.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults":

"nextToken": To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

"status": The status of the devices in the import task.

Main.Iot_Wireless.list_event_configurations — Method

list_event_configurations(resource_type)
list_event_configurations(resource_type, params::Dict{String,<:Any})

List event configurations where at least one event topic has been enabled.

Arguments

resource_type: Resource type to filter event configurations.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults":

"nextToken": To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

Main.Iot_Wireless.list_fuota_tasks — Method

list_fuota_tasks()
list_fuota_tasks(params::Dict{String,<:Any})

Lists the FUOTA tasks registered to your AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults":

"nextToken": To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

Main.Iot_Wireless.list_multicast_groups — Method

list_multicast_groups()
list_multicast_groups(params::Dict{String,<:Any})

Lists the multicast groups registered to your AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults":

"nextToken": To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

Main.Iot_Wireless.list_multicast_groups_by_fuota_task — Method

list_multicast_groups_by_fuota_task(id)
list_multicast_groups_by_fuota_task(id, params::Dict{String,<:Any})

List all multicast groups associated with a fuota task.

Arguments

id:

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults":

"nextToken": To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

Main.Iot_Wireless.list_network_analyzer_configurations — Method

list_network_analyzer_configurations()
list_network_analyzer_configurations(params::Dict{String,<:Any})

Lists the network analyzer configurations.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults":

"nextToken": To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

Main.Iot_Wireless.list_partner_accounts — Method

list_partner_accounts()
list_partner_accounts(params::Dict{String,<:Any})

Lists the partner accounts associated with your AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return in this operation.

"nextToken": To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

Main.Iot_Wireless.list_position_configurations — Method

list_position_configurations()
list_position_configurations(params::Dict{String,<:Any})

List position configurations for a given resource, such as positioning solvers. This action is no longer supported. Calls to retrieve position information should use the GetResourcePosition API operation instead.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults":

"nextToken": To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

"resourceType": Resource type for which position configurations are listed.

Main.Iot_Wireless.list_queued_messages — Method

list_queued_messages(id)
list_queued_messages(id, params::Dict{String,<:Any})

List queued messages in the downlink queue.

Arguments

id: The ID of a given wireless device which the downlink message packets are being sent.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"WirelessDeviceType": The wireless device type, whic can be either Sidewalk or LoRaWAN.

"maxResults": The maximum number of results to return in this operation.

"nextToken": To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

Main.Iot_Wireless.list_service_profiles — Method

list_service_profiles()
list_service_profiles(params::Dict{String,<:Any})

Lists the service profiles registered to your AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return in this operation.

"nextToken": To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

Main.Iot_Wireless.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists the tags (metadata) you have assigned to the resource.

Arguments

resource_arn: The ARN of the resource for which you want to list tags.

Main.Iot_Wireless.list_wireless_device_import_tasks — Method

list_wireless_device_import_tasks()
list_wireless_device_import_tasks(params::Dict{String,<:Any})

List wireless devices that have been added to an import task.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults":

"nextToken": To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

Main.Iot_Wireless.list_wireless_devices — Method

list_wireless_devices()
list_wireless_devices(params::Dict{String,<:Any})

Lists the wireless devices registered to your AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"destinationName": A filter to list only the wireless devices that use this destination.

"deviceProfileId": A filter to list only the wireless devices that use this device profile.

"fuotaTaskId":

"maxResults": The maximum number of results to return in this operation.

"multicastGroupId":

"nextToken": To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

"serviceProfileId": A filter to list only the wireless devices that use this service profile.

"wirelessDeviceType": A filter to list only the wireless devices that use this wireless device type.

Main.Iot_Wireless.list_wireless_gateway_task_definitions — Method

list_wireless_gateway_task_definitions()
list_wireless_gateway_task_definitions(params::Dict{String,<:Any})

List the wireless gateway tasks definitions registered to your AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return in this operation.

"nextToken": To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

"taskDefinitionType": A filter to list only the wireless gateway task definitions that use this task definition type.

Main.Iot_Wireless.list_wireless_gateways — Method

list_wireless_gateways()
list_wireless_gateways(params::Dict{String,<:Any})

Lists the wireless gateways registered to your AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return in this operation.

"nextToken": To retrieve the next set of results, the nextToken value from a previous response; otherwise null to receive the first set of results.

Main.Iot_Wireless.put_position_configuration — Method

put_position_configuration(resource_identifier, resource_type)
put_position_configuration(resource_identifier, resource_type, params::Dict{String,<:Any})

Put position configuration for a given resource. This action is no longer supported. Calls to update the position configuration should use the UpdateResourcePosition API operation instead.

Arguments

resource_identifier: Resource identifier used to update the position configuration.

resource_type: Resource type of the resource for which you want to update the position configuration.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Destination": The position data destination that describes the AWS IoT rule that processes the device's position data for use by AWS IoT Core for LoRaWAN.

"Solvers": The positioning solvers used to update the position configuration of the resource.

Main.Iot_Wireless.put_resource_log_level — Method

put_resource_log_level(log_level, resource_identifier, resource_type)
put_resource_log_level(log_level, resource_identifier, resource_type, params::Dict{String,<:Any})

Sets the log-level override for a resource-ID and resource-type. This option can be specified for a wireless gateway or a wireless device. A limit of 200 log level override can be set per account.

Arguments

log_level:

resource_identifier:

resource_type: The type of the resource, which can be WirelessDevice or WirelessGateway.

Main.Iot_Wireless.reset_all_resource_log_levels — Method

reset_all_resource_log_levels()
reset_all_resource_log_levels(params::Dict{String,<:Any})

Removes the log-level overrides for all resources; both wireless devices and wireless gateways.

Main.Iot_Wireless.reset_resource_log_level — Method

reset_resource_log_level(resource_identifier, resource_type)
reset_resource_log_level(resource_identifier, resource_type, params::Dict{String,<:Any})

Removes the log-level override, if any, for a specific resource-ID and resource-type. It can be used for a wireless device or a wireless gateway.

Arguments

resource_identifier:

resource_type: The type of the resource, which can be WirelessDevice or WirelessGateway.

Main.Iot_Wireless.send_data_to_multicast_group — Method

send_data_to_multicast_group(id, payload_data, wireless_metadata)
send_data_to_multicast_group(id, payload_data, wireless_metadata, params::Dict{String,<:Any})

Sends the specified data to a multicast group.

Arguments

id:

payload_data:

wireless_metadata:

Main.Iot_Wireless.send_data_to_wireless_device — Method

send_data_to_wireless_device(id, payload_data, transmit_mode)
send_data_to_wireless_device(id, payload_data, transmit_mode, params::Dict{String,<:Any})

Sends a decrypted application data frame to a device.

Arguments

id: The ID of the wireless device to receive the data.

payload_data:

transmit_mode: The transmit mode to use to send data to the wireless device. Can be: 0 for UM (unacknowledge mode) or 1 for AM (acknowledge mode).

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"WirelessMetadata": Metadata about the message request.

Main.Iot_Wireless.start_bulk_associate_wireless_device_with_multicast_group — Method

start_bulk_associate_wireless_device_with_multicast_group(id)
start_bulk_associate_wireless_device_with_multicast_group(id, params::Dict{String,<:Any})

Starts a bulk association of all qualifying wireless devices with a multicast group.

Arguments

id:

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"QueryString":

"Tags":

Main.Iot_Wireless.start_bulk_disassociate_wireless_device_from_multicast_group — Method

start_bulk_disassociate_wireless_device_from_multicast_group(id)
start_bulk_disassociate_wireless_device_from_multicast_group(id, params::Dict{String,<:Any})

Starts a bulk disassociatin of all qualifying wireless devices from a multicast group.

Arguments

id:

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"QueryString":

"Tags":

Main.Iot_Wireless.start_fuota_task — Method

start_fuota_task(id)
start_fuota_task(id, params::Dict{String,<:Any})

Starts a FUOTA task.

Arguments

id:

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"LoRaWAN":

Main.Iot_Wireless.start_multicast_group_session — Method

start_multicast_group_session(id, lo_ra_wan)
start_multicast_group_session(id, lo_ra_wan, params::Dict{String,<:Any})

Starts a multicast group session.

Arguments

id:

lo_ra_wan:

Main.Iot_Wireless.start_single_wireless_device_import_task — Method

start_single_wireless_device_import_task(destination_name, sidewalk)
start_single_wireless_device_import_task(destination_name, sidewalk, params::Dict{String,<:Any})

Start import task for a single wireless device.

Arguments

destination_name: The name of the Sidewalk destination that describes the IoT rule to route messages from the device in the import task that will be onboarded to AWS IoT Wireless.

sidewalk: The Sidewalk-related parameters for importing a single wireless device.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientRequestToken":

"DeviceName": The name of the wireless device for which an import task is being started.

"Tags":

Main.Iot_Wireless.start_wireless_device_import_task — Method

start_wireless_device_import_task(destination_name, sidewalk)
start_wireless_device_import_task(destination_name, sidewalk, params::Dict{String,<:Any})

Start import task for provisioning Sidewalk devices in bulk using an S3 CSV file.

Arguments

destination_name: The name of the Sidewalk destination that describes the IoT rule to route messages from the devices in the import task that are onboarded to AWS IoT Wireless.

sidewalk: The Sidewalk-related parameters for importing wireless devices that need to be provisioned in bulk.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ClientRequestToken":

"Tags":

Main.Iot_Wireless.tag_resource — Method

tag_resource(tags, resource_arn)
tag_resource(tags, resource_arn, params::Dict{String,<:Any})

Adds a tag to a resource.

Arguments

tags: Adds to or modifies the tags of the given resource. Tags are metadata that you can use to manage a resource.

resource_arn: The ARN of the resource to add tags to.

Main.Iot_Wireless.test_wireless_device — Method

test_wireless_device(id)
test_wireless_device(id, params::Dict{String,<:Any})

Simulates a provisioned device by sending an uplink data payload of Hello.

Arguments

id: The ID of the wireless device to test.

Main.Iot_Wireless.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes one or more tags from a resource.

Arguments

resource_arn: The ARN of the resource to remove tags from.

tag_keys: A list of the keys of the tags to remove from the resource.

Main.Iot_Wireless.update_destination — Method

update_destination(name)
update_destination(name, params::Dict{String,<:Any})

Updates properties of a destination.

Arguments

name: The new name of the resource.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A new description of the resource.

"Expression": The new rule name or topic rule to send messages to.

"ExpressionType": The type of value in Expression.

"RoleArn": The ARN of the IAM Role that authorizes the destination.

Main.Iot_Wireless.update_event_configuration_by_resource_types — Method

update_event_configuration_by_resource_types()
update_event_configuration_by_resource_types(params::Dict{String,<:Any})

Update the event configuration based on resource types.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ConnectionStatus": Connection status resource type event configuration object for enabling and disabling wireless gateway topic.

"DeviceRegistrationState": Device registration state resource type event configuration object for enabling and disabling wireless gateway topic.

"Join": Join resource type event configuration object for enabling and disabling wireless device topic.

"MessageDeliveryStatus": Message delivery status resource type event configuration object for enabling and disabling wireless device topic.

"Proximity": Proximity resource type event configuration object for enabling and disabling wireless gateway topic.

Main.Iot_Wireless.update_fuota_task — Method

update_fuota_task(id)
update_fuota_task(id, params::Dict{String,<:Any})

Updates properties of a FUOTA task.

Arguments

id:

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description":

"FirmwareUpdateImage":

"FirmwareUpdateRole":

"FragmentIntervalMS":

"FragmentSizeBytes":

"LoRaWAN":

"Name":

"RedundancyPercent":

Main.Iot_Wireless.update_log_levels_by_resource_types — Method

update_log_levels_by_resource_types()
update_log_levels_by_resource_types(params::Dict{String,<:Any})

Set default log level, or log levels by resource types. This can be for wireless device log options or wireless gateways log options and is used to control the log messages that'll be displayed in CloudWatch.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DefaultLogLevel":

"WirelessDeviceLogOptions":

"WirelessGatewayLogOptions":

Main.Iot_Wireless.update_multicast_group — Method

update_multicast_group(id)
update_multicast_group(id, params::Dict{String,<:Any})

Updates properties of a multicast group session.

Arguments

id:

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description":

"LoRaWAN":

"Name":

Main.Iot_Wireless.update_network_analyzer_configuration — Method

update_network_analyzer_configuration(configuration_name)
update_network_analyzer_configuration(configuration_name, params::Dict{String,<:Any})

Update network analyzer configuration.

Arguments

configuration_name:

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description":

"MulticastGroupsToAdd": Multicast group resources to add to the network analyzer configuration. Provide the MulticastGroupId of the resource to add in the input array.

"MulticastGroupsToRemove": Multicast group resources to remove from the network analyzer configuration. Provide the MulticastGroupId of the resource to remove in the input array.

"TraceContent":

"WirelessDevicesToAdd": Wireless device resources to add to the network analyzer configuration. Provide the WirelessDeviceId of the resource to add in the input array.

"WirelessDevicesToRemove": Wireless device resources to remove from the network analyzer configuration. Provide the WirelessDeviceId of the resources to remove in the input array.

"WirelessGatewaysToAdd": Wireless gateway resources to add to the network analyzer configuration. Provide the WirelessGatewayId of the resource to add in the input array.

"WirelessGatewaysToRemove": Wireless gateway resources to remove from the network analyzer configuration. Provide the WirelessGatewayId of the resources to remove in the input array.

Main.Iot_Wireless.update_partner_account — Method

update_partner_account(partner_account_id, sidewalk, partner_type)
update_partner_account(partner_account_id, sidewalk, partner_type, params::Dict{String,<:Any})

Updates properties of a partner account.

Arguments

partner_account_id: The ID of the partner account to update.

sidewalk: The Sidewalk account credentials.

partner_type: The partner type.

Main.Iot_Wireless.update_position — Method

update_position(position, resource_identifier, resource_type)
update_position(position, resource_identifier, resource_type, params::Dict{String,<:Any})

Update the position information of a resource. This action is no longer supported. Calls to update the position information should use the UpdateResourcePosition API operation instead.

Arguments

position: The position information of the resource.

resource_identifier: Resource identifier of the resource for which position is updated.

resource_type: Resource type of the resource for which position is updated.

Main.Iot_Wireless.update_resource_event_configuration — Method

update_resource_event_configuration(identifier, identifier_type)
update_resource_event_configuration(identifier, identifier_type, params::Dict{String,<:Any})

Update the event configuration for a particular resource identifier.

Arguments

identifier: Resource identifier to opt in for event messaging.

identifier_type: Identifier type of the particular resource identifier for event configuration.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"ConnectionStatus": Event configuration for the connection status event.

"DeviceRegistrationState": Event configuration for the device registration state event.

"Join": Event configuration for the join event.

"MessageDeliveryStatus": Event configuration for the message delivery status event.

"Proximity": Event configuration for the proximity event.

"partnerType": Partner type of the resource if the identifier type is PartnerAccountId

Main.Iot_Wireless.update_resource_position — Method

update_resource_position(resource_identifier, resource_type)
update_resource_position(resource_identifier, resource_type, params::Dict{String,<:Any})

Update the position information of a given wireless device or a wireless gateway resource. The position coordinates are based on the World Geodetic System (WGS84).

Arguments

resource_identifier: The identifier of the resource for which position information is updated. It can be the wireless device ID or the wireless gateway ID, depending on the resource type.

resource_type: The type of resource for which position information is updated, which can be a wireless device or a wireless gateway.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"GeoJsonPayload": The position information of the resource, displayed as a JSON payload. The payload uses the GeoJSON format, which a format that's used to encode geographic data structures. For more information, see GeoJSON.

Main.Iot_Wireless.update_wireless_device — Method

update_wireless_device(id)
update_wireless_device(id, params::Dict{String,<:Any})

Updates properties of a wireless device.

Arguments

id: The ID of the resource to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A new description of the resource.

"DestinationName": The name of the new destination for the device.

"LoRaWAN": The updated wireless device's configuration.

"Name": The new name of the resource.

"Positioning": FPort values for the GNSS, stream, and ClockSync functions of the positioning information.

Main.Iot_Wireless.update_wireless_device_import_task — Method

update_wireless_device_import_task(id, sidewalk)
update_wireless_device_import_task(id, sidewalk, params::Dict{String,<:Any})

Update an import task to add more devices to the task.

Arguments

id: The identifier of the import task to be updated.

sidewalk: The Sidewalk-related parameters of the import task to be updated.

Main.Iot_Wireless.update_wireless_gateway — Method

update_wireless_gateway(id)
update_wireless_gateway(id, params::Dict{String,<:Any})

Updates properties of a wireless gateway.

Arguments

id: The ID of the resource to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": A new description of the resource.

"JoinEuiFilters":

"MaxEirp": The MaxEIRP value.

"Name": The new name of the resource.

"NetIdFilters":