Iottwinmaker

This page documents function available when using the Iottwinmaker module, created with @service Iottwinmaker.

Index

Main.Iottwinmaker.batch_put_property_values

Main.Iottwinmaker.create_component_type

Main.Iottwinmaker.create_entity

Main.Iottwinmaker.create_scene

Main.Iottwinmaker.create_sync_job

Main.Iottwinmaker.create_workspace

Main.Iottwinmaker.delete_component_type

Main.Iottwinmaker.delete_entity

Main.Iottwinmaker.delete_scene

Main.Iottwinmaker.delete_sync_job

Main.Iottwinmaker.delete_workspace

Main.Iottwinmaker.execute_query

Main.Iottwinmaker.get_component_type

Main.Iottwinmaker.get_entity

Main.Iottwinmaker.get_pricing_plan

Main.Iottwinmaker.get_property_value

Main.Iottwinmaker.get_property_value_history

Main.Iottwinmaker.get_scene

Main.Iottwinmaker.get_sync_job

Main.Iottwinmaker.get_workspace

Main.Iottwinmaker.list_component_types

Main.Iottwinmaker.list_entities

Main.Iottwinmaker.list_scenes

Main.Iottwinmaker.list_sync_jobs

Main.Iottwinmaker.list_sync_resources

Main.Iottwinmaker.list_tags_for_resource

Main.Iottwinmaker.list_workspaces

Main.Iottwinmaker.tag_resource

Main.Iottwinmaker.untag_resource

Main.Iottwinmaker.update_component_type

Main.Iottwinmaker.update_entity

Main.Iottwinmaker.update_pricing_plan

Main.Iottwinmaker.update_scene

Main.Iottwinmaker.update_workspace

Documentation

Main.Iottwinmaker.batch_put_property_values — Method

batch_put_property_values(entries, workspace_id)
batch_put_property_values(entries, workspace_id, params::Dict{String,<:Any})

Sets values for multiple time series properties.

Arguments

entries: An object that maps strings to the property value entries to set. Each string in the mapping must be unique to this object.

workspace_id: The ID of the workspace that contains the properties to set.

Main.Iottwinmaker.create_component_type — Method

create_component_type(component_type_id, workspace_id)
create_component_type(component_type_id, workspace_id, params::Dict{String,<:Any})

Creates a component type.

Arguments

component_type_id: The ID of the component type.

workspace_id: The ID of the workspace that contains the component type.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"componentTypeName": A friendly name for the component type.

"description": The description of the component type.

"extendsFrom": Specifies the parent component type to extend.

"functions": An object that maps strings to the functions in the component type. Each string in the mapping must be unique to this object.

"isSingleton": A Boolean value that specifies whether an entity can have more than one component of this type.

"propertyDefinitions": An object that maps strings to the property definitions in the component type. Each string in the mapping must be unique to this object.

"propertyGroups":

"tags": Metadata that you can use to manage the component type.

Main.Iottwinmaker.create_entity — Method

create_entity(entity_name, workspace_id)
create_entity(entity_name, workspace_id, params::Dict{String,<:Any})

Creates an entity.

Arguments

entity_name: The name of the entity.

workspace_id: The ID of the workspace that contains the entity.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"components": An object that maps strings to the components in the entity. Each string in the mapping must be unique to this object.

"description": The description of the entity.

"entityId": The ID of the entity.

"parentEntityId": The ID of the entity's parent entity.

"tags": Metadata that you can use to manage the entity.

Main.Iottwinmaker.create_scene — Method

create_scene(content_location, scene_id, workspace_id)
create_scene(content_location, scene_id, workspace_id, params::Dict{String,<:Any})

Creates a scene.

Arguments

content_location: The relative path that specifies the location of the content definition file.

scene_id: The ID of the scene.

workspace_id: The ID of the workspace that contains the scene.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"capabilities": A list of capabilities that the scene uses to render itself.

"description": The description for this scene.

"sceneMetadata": The request metadata.

"tags": Metadata that you can use to manage the scene.

Main.Iottwinmaker.create_sync_job — Method

create_sync_job(sync_role, sync_source, workspace_id)
create_sync_job(sync_role, sync_source, workspace_id, params::Dict{String,<:Any})

This action creates a SyncJob.

Arguments

sync_role: The SyncJob IAM role. This IAM role is used by the SyncJob to read from the syncSource, and create, update, or delete the corresponding resources.

sync_source: The sync source. Currently the only supported syncSoource is SITEWISE .

workspace_id: The workspace ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"tags": The SyncJob tags.

Main.Iottwinmaker.create_workspace — Method

create_workspace(role, s3_location, workspace_id)
create_workspace(role, s3_location, workspace_id, params::Dict{String,<:Any})

Creates a workplace.

Arguments

role: The ARN of the execution role associated with the workspace.

s3_location: The ARN of the S3 bucket where resources associated with the workspace are stored.

workspace_id: The ID of the workspace.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The description of the workspace.

"tags": Metadata that you can use to manage the workspace

Main.Iottwinmaker.delete_component_type — Method

delete_component_type(component_type_id, workspace_id)
delete_component_type(component_type_id, workspace_id, params::Dict{String,<:Any})

Deletes a component type.

Arguments

component_type_id: The ID of the component type to delete.

workspace_id: The ID of the workspace that contains the component type.

Main.Iottwinmaker.delete_entity — Method

delete_entity(entity_id, workspace_id)
delete_entity(entity_id, workspace_id, params::Dict{String,<:Any})

Deletes an entity.

Arguments

entity_id: The ID of the entity to delete.

workspace_id: The ID of the workspace that contains the entity to delete.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"isRecursive": A Boolean value that specifies whether the operation deletes child entities.

Main.Iottwinmaker.delete_scene — Method

delete_scene(scene_id, workspace_id)
delete_scene(scene_id, workspace_id, params::Dict{String,<:Any})

Deletes a scene.

Arguments

scene_id: The ID of the scene to delete.

workspace_id: The ID of the workspace.

Main.Iottwinmaker.delete_sync_job — Method

delete_sync_job(sync_source, workspace_id)
delete_sync_job(sync_source, workspace_id, params::Dict{String,<:Any})

Delete the SyncJob.

Arguments

sync_source: The sync source. Currently the only supported syncSource is SITEWISE .

workspace_id: The workspace ID.

Main.Iottwinmaker.delete_workspace — Method

delete_workspace(workspace_id)
delete_workspace(workspace_id, params::Dict{String,<:Any})

Deletes a workspace.

Arguments

workspace_id: The ID of the workspace to delete.

Main.Iottwinmaker.execute_query — Method

execute_query(query_statement, workspace_id)
execute_query(query_statement, workspace_id, params::Dict{String,<:Any})

Run queries to access information from your knowledge graph of entities within individual workspaces.

Arguments

query_statement: The query statement.

workspace_id: The ID of the workspace.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return at one time. The default is 25. Valid Range: Minimum value of 1. Maximum value of 250.

"nextToken": The string that specifies the next page of results.

Main.Iottwinmaker.get_component_type — Method

get_component_type(component_type_id, workspace_id)
get_component_type(component_type_id, workspace_id, params::Dict{String,<:Any})

Retrieves information about a component type.

Arguments

component_type_id: The ID of the component type.

workspace_id: The ID of the workspace that contains the component type.

Main.Iottwinmaker.get_entity — Method

get_entity(entity_id, workspace_id)
get_entity(entity_id, workspace_id, params::Dict{String,<:Any})

Retrieves information about an entity.

Arguments

entity_id: The ID of the entity.

workspace_id: The ID of the workspace.

Main.Iottwinmaker.get_pricing_plan — Method

get_pricing_plan()
get_pricing_plan(params::Dict{String,<:Any})

Gets the pricing plan.

Main.Iottwinmaker.get_property_value — Method

get_property_value(selected_properties, workspace_id)
get_property_value(selected_properties, workspace_id, params::Dict{String,<:Any})

Gets the property values for a component, component type, entity, or workspace. You must specify a value for either componentName, componentTypeId, entityId, or workspaceId.

Arguments

selected_properties: The properties whose values the operation returns.

workspace_id: The ID of the workspace whose values the operation returns.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"componentName": The name of the component whose property values the operation returns.

"componentTypeId": The ID of the component type whose property values the operation returns.

"entityId": The ID of the entity whose property values the operation returns.

"maxResults": The maximum number of results to return at one time. The default is 25. Valid Range: Minimum value of 1. Maximum value of 250.

"nextToken": The string that specifies the next page of results.

"propertyGroupName": The property group name.

"tabularConditions": The tabular conditions.

Main.Iottwinmaker.get_property_value_history — Method

get_property_value_history(selected_properties, workspace_id)
get_property_value_history(selected_properties, workspace_id, params::Dict{String,<:Any})

Retrieves information about the history of a time series property value for a component, component type, entity, or workspace. You must specify a value for workspaceId. For entity-specific queries, specify values for componentName and entityId. For cross-entity quries, specify a value for componentTypeId.

Arguments

selected_properties: A list of properties whose value histories the request retrieves.

workspace_id: The ID of the workspace.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"componentName": The name of the component.

"componentTypeId": The ID of the component type.

"endDateTime": The date and time of the latest property value to return.

"endTime": The ISO8601 DateTime of the latest property value to return. For more information about the ISO8601 DateTime format, see the data type PropertyValue.

"entityId": The ID of the entity.

"interpolation": An object that specifies the interpolation type and the interval over which to interpolate data.

"maxResults": The maximum number of results to return at one time. The default is 25. Valid Range: Minimum value of 1. Maximum value of 250.

"nextToken": The string that specifies the next page of results.

"orderByTime": The time direction to use in the result order.

"propertyFilters": A list of objects that filter the property value history request.

"startDateTime": The date and time of the earliest property value to return.

"startTime": The ISO8601 DateTime of the earliest property value to return. For more information about the ISO8601 DateTime format, see the data type PropertyValue.

Main.Iottwinmaker.get_scene — Method

get_scene(scene_id, workspace_id)
get_scene(scene_id, workspace_id, params::Dict{String,<:Any})

Retrieves information about a scene.

Arguments

scene_id: The ID of the scene.

workspace_id: The ID of the workspace that contains the scene.

Main.Iottwinmaker.get_sync_job — Method

get_sync_job(sync_source)
get_sync_job(sync_source, params::Dict{String,<:Any})

Gets the SyncJob.

Arguments

sync_source: The sync source. Currently the only supported syncSource is SITEWISE .

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"workspace": The workspace ID.

Main.Iottwinmaker.get_workspace — Method

get_workspace(workspace_id)
get_workspace(workspace_id, params::Dict{String,<:Any})

Retrieves information about a workspace.

Arguments

workspace_id: The ID of the workspace.

Main.Iottwinmaker.list_component_types — Method

list_component_types(workspace_id)
list_component_types(workspace_id, params::Dict{String,<:Any})

Lists all component types in a workspace.

Arguments

workspace_id: The ID of the workspace.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filters": A list of objects that filter the request.

"maxResults": The maximum number of results to return at one time. The default is 25. Valid Range: Minimum value of 1. Maximum value of 250.

"nextToken": The string that specifies the next page of results.

Main.Iottwinmaker.list_entities — Method

list_entities(workspace_id)
list_entities(workspace_id, params::Dict{String,<:Any})

Lists all entities in a workspace.

Arguments

workspace_id: The ID of the workspace.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filters": A list of objects that filter the request. Only one object is accepted as a valid input.

"maxResults": The maximum number of results to return at one time. The default is 25. Valid Range: Minimum value of 1. Maximum value of 250.

"nextToken": The string that specifies the next page of results.

Main.Iottwinmaker.list_scenes — Method

list_scenes(workspace_id)
list_scenes(workspace_id, params::Dict{String,<:Any})

Lists all scenes in a workspace.

Arguments

workspace_id: The ID of the workspace that contains the scenes.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Specifies the maximum number of results to display.

"nextToken": The string that specifies the next page of results.

Main.Iottwinmaker.list_sync_jobs — Method

list_sync_jobs(workspace_id)
list_sync_jobs(workspace_id, params::Dict{String,<:Any})

List all SyncJobs.

Arguments

workspace_id: The ID of the workspace that contains the sync job.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return at one time. The default is 50. Valid Range: Minimum value of 0. Maximum value of 200.

"nextToken": The string that specifies the next page of results.

Main.Iottwinmaker.list_sync_resources — Method

list_sync_resources(sync_source, workspace_id)
list_sync_resources(sync_source, workspace_id, params::Dict{String,<:Any})

Lists the sync resources.

Arguments

sync_source: The sync source. Currently the only supported syncSource is SITEWISE .

workspace_id: The ID of the workspace that contains the sync job.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filters": A list of objects that filter the request. The following filter combinations are supported: Filter with state Filter with ResourceType and ResourceId Filter with ResourceType and ExternalId

"maxResults": The maximum number of results to return at one time. The default is 50. Valid Range: Minimum value of 0. Maximum value of 200.

"nextToken": The string that specifies the next page of results.

Main.Iottwinmaker.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists all tags associated with a resource.

Arguments

resource_arn: The ARN of the resource.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return at one time. The default is 25. Valid Range: Minimum value of 1. Maximum value of 250.

"nextToken": The string that specifies the next page of results.

Main.Iottwinmaker.list_workspaces — Method

list_workspaces()
list_workspaces(params::Dict{String,<:Any})

Retrieves information about workspaces in the current account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return at one time. The default is 25. Valid Range: Minimum value of 1. Maximum value of 250.

"nextToken": The string that specifies the next page of results.

Main.Iottwinmaker.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Adds tags to a resource.

Arguments

resource_arn: The ARN of the resource.

tags: Metadata to add to this resource.

Main.Iottwinmaker.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes tags from a resource.

Arguments

resource_arn: The ARN of the resource.

tag_keys: A list of tag key names to remove from the resource. You don't specify the value. Both the key and its associated value are removed.

Main.Iottwinmaker.update_component_type — Method

update_component_type(component_type_id, workspace_id)
update_component_type(component_type_id, workspace_id, params::Dict{String,<:Any})

Updates information in a component type.

Arguments

component_type_id: The ID of the component type.

workspace_id: The ID of the workspace.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"componentTypeName": The component type name.

"description": The description of the component type.

"extendsFrom": Specifies the component type that this component type extends.

"functions": An object that maps strings to the functions in the component type. Each string in the mapping must be unique to this object.

"isSingleton": A Boolean value that specifies whether an entity can have more than one component of this type.

"propertyDefinitions": An object that maps strings to the property definitions in the component type. Each string in the mapping must be unique to this object.

"propertyGroups": The property groups.

Main.Iottwinmaker.update_entity — Method

update_entity(entity_id, workspace_id)
update_entity(entity_id, workspace_id, params::Dict{String,<:Any})

Updates an entity.

Arguments

entity_id: The ID of the entity.

workspace_id: The ID of the workspace that contains the entity.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"componentUpdates": An object that maps strings to the component updates in the request. Each string in the mapping must be unique to this object.

"description": The description of the entity.

"entityName": The name of the entity.

"parentEntityUpdate": An object that describes the update request for a parent entity.

Main.Iottwinmaker.update_pricing_plan — Method

update_pricing_plan(pricing_mode)
update_pricing_plan(pricing_mode, params::Dict{String,<:Any})

Update the pricing plan.

Arguments

pricing_mode: The pricing mode.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"bundleNames": The bundle names.

Main.Iottwinmaker.update_scene — Method

update_scene(scene_id, workspace_id)
update_scene(scene_id, workspace_id, params::Dict{String,<:Any})

Updates a scene.

Arguments

scene_id: The ID of the scene.

workspace_id: The ID of the workspace that contains the scene.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"capabilities": A list of capabilities that the scene uses to render.

"contentLocation": The relative path that specifies the location of the content definition file.

"description": The description of this scene.

"sceneMetadata": The scene metadata.

Main.Iottwinmaker.update_workspace — Method

update_workspace(workspace_id)
update_workspace(workspace_id, params::Dict{String,<:Any})

Updates a workspace.

Arguments

workspace_id: The ID of the workspace.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": The description of the workspace.

"role": The ARN of the execution role associated with the workspace.