Location

This page documents function available when using the Location module, created with @service Location.

Index

Main.Location.associate_tracker_consumer

Main.Location.batch_delete_device_position_history

Main.Location.batch_delete_geofence

Main.Location.batch_evaluate_geofences

Main.Location.batch_get_device_position

Main.Location.batch_put_geofence

Main.Location.batch_update_device_position

Main.Location.calculate_route

Main.Location.create_geofence_collection

Main.Location.create_map

Main.Location.create_place_index

Main.Location.create_route_calculator

Main.Location.create_tracker

Main.Location.delete_geofence_collection

Main.Location.delete_map

Main.Location.delete_place_index

Main.Location.delete_route_calculator

Main.Location.delete_tracker

Main.Location.describe_geofence_collection

Main.Location.describe_map

Main.Location.describe_place_index

Main.Location.describe_route_calculator

Main.Location.describe_tracker

Main.Location.disassociate_tracker_consumer

Main.Location.get_device_position

Main.Location.get_device_position_history

Main.Location.get_geofence

Main.Location.get_map_glyphs

Main.Location.get_map_sprites

Main.Location.get_map_style_descriptor

Main.Location.get_map_tile

Main.Location.list_device_positions

Main.Location.list_geofence_collections

Main.Location.list_geofences

Main.Location.list_maps

Main.Location.list_place_indexes

Main.Location.list_route_calculators

Main.Location.list_tags_for_resource

Main.Location.list_tracker_consumers

Main.Location.list_trackers

Main.Location.put_geofence

Main.Location.search_place_index_for_position

Main.Location.search_place_index_for_suggestions

Main.Location.search_place_index_for_text

Main.Location.tag_resource

Main.Location.untag_resource

Main.Location.update_geofence_collection

Main.Location.update_map

Main.Location.update_place_index

Main.Location.update_route_calculator

Main.Location.update_tracker

Documentation

Main.Location.associate_tracker_consumer — Method

associate_tracker_consumer(consumer_arn, tracker_name)
associate_tracker_consumer(consumer_arn, tracker_name, params::Dict{String,<:Any})

Creates an association between a geofence collection and a tracker resource. This allows the tracker resource to communicate location data to the linked geofence collection. You can associate up to five geofence collections to each tracker resource. Currently not supported — Cross-account configurations, such as creating associations between a tracker resource in one account and a geofence collection in another account.

Arguments

consumer_arn: The Amazon Resource Name (ARN) for the geofence collection to be associated to tracker resource. Used when you need to specify a resource across all AWS. Format example: arn:aws:geo:region:account-id:geofence-collection/ExampleGeofenceCollectionConsumer

tracker_name: The name of the tracker resource to be associated with a geofence collection.

Main.Location.batch_delete_device_position_history — Method

batch_delete_device_position_history(device_ids, tracker_name)
batch_delete_device_position_history(device_ids, tracker_name, params::Dict{String,<:Any})

Deletes the position history of one or more devices from a tracker resource.

Arguments

device_ids: Devices whose position history you want to delete. For example, for two devices: “DeviceIds” : [DeviceId1,DeviceId2]

tracker_name: The name of the tracker resource to delete the device position history from.

Main.Location.batch_delete_geofence — Method

batch_delete_geofence(collection_name, geofence_ids)
batch_delete_geofence(collection_name, geofence_ids, params::Dict{String,<:Any})

Deletes a batch of geofences from a geofence collection. This operation deletes the resource permanently.

Arguments

collection_name: The geofence collection storing the geofences to be deleted.

geofence_ids: The batch of geofences to be deleted.

Main.Location.batch_evaluate_geofences — Method

batch_evaluate_geofences(collection_name, device_position_updates)
batch_evaluate_geofences(collection_name, device_position_updates, params::Dict{String,<:Any})

Evaluates device positions against the geofence geometries from a given geofence collection. This operation always returns an empty response because geofences are asynchronously evaluated. The evaluation determines if the device has entered or exited a geofenced area, and then publishes one of the following events to Amazon EventBridge: ENTER if Amazon Location determines that the tracked device has entered a geofenced area. EXIT if Amazon Location determines that the tracked device has exited a geofenced area. The last geofence that a device was observed within is tracked for 30 days after the most recent device position update. Geofence evaluation uses the given device position. It does not account for the optional Accuracy of a DevicePositionUpdate.

Arguments

collection_name: The geofence collection used in evaluating the position of devices against its geofences.

device_position_updates: Contains device details for each device to be evaluated against the given geofence collection.

Main.Location.batch_get_device_position — Method

batch_get_device_position(device_ids, tracker_name)
batch_get_device_position(device_ids, tracker_name, params::Dict{String,<:Any})

Lists the latest device positions for requested devices.

Arguments

device_ids: Devices whose position you want to retrieve. For example, for two devices: device-ids=DeviceId1&amp;device-ids=DeviceId2

tracker_name: The tracker resource retrieving the device position.

Main.Location.batch_put_geofence — Method

batch_put_geofence(collection_name, entries)
batch_put_geofence(collection_name, entries, params::Dict{String,<:Any})

A batch request for storing geofence geometries into a given geofence collection, or updates the geometry of an existing geofence if a geofence ID is included in the request.

Arguments

collection_name: The geofence collection storing the geofences.

entries: The batch of geofences to be stored in a geofence collection.

Main.Location.batch_update_device_position — Method

batch_update_device_position(tracker_name, updates)
batch_update_device_position(tracker_name, updates, params::Dict{String,<:Any})

Uploads position update data for one or more devices to a tracker resource. Amazon Location uses the data when it reports the last known device position and position history. Amazon Location retains location data for 30 days. Position updates are handled based on the PositionFiltering property of the tracker. When PositionFiltering is set to TimeBased, updates are evaluated against linked geofence collections, and location data is stored at a maximum of one position per 30 second interval. If your update frequency is more often than every 30 seconds, only one update per 30 seconds is stored for each unique device ID. When PositionFiltering is set to DistanceBased filtering, location data is stored and evaluated against linked geofence collections only if the device has moved more than 30 m (98.4 ft). When PositionFiltering is set to AccuracyBased filtering, location data is stored and evaluated against linked geofence collections only if the device has moved more than the measured accuracy. For example, if two consecutive updates from a device have a horizontal accuracy of 5 m and 10 m, the second update is neither stored or evaluated if the device has moved less than 15 m. If PositionFiltering is set to AccuracyBased filtering, Amazon Location uses the default value { "Horizontal": 0} when accuracy is not provided on a DevicePositionUpdate.

Arguments

tracker_name: The name of the tracker resource to update.

updates: Contains the position update details for each device.

Main.Location.calculate_route — Method

calculate_route(calculator_name, departure_position, destination_position)
calculate_route(calculator_name, departure_position, destination_position, params::Dict{String,<:Any})

Calculates a route given the following required parameters: DeparturePostiton and DestinationPosition. Requires that you first create a route calculator resource. By default, a request that doesn't specify a departure time uses the best time of day to travel with the best traffic conditions when calculating the route. Additional options include: Specifying a departure time using either DepartureTime or DepartureNow. This calculates a route based on predictive traffic data at the given time. You can't specify both DepartureTime and DepartureNow in a single request. Specifying both parameters returns a validation error. Specifying a travel mode using TravelMode. This lets you specify an additional route preference such as CarModeOptions if traveling by Car, or TruckModeOptions if traveling by Truck.

Arguments

calculator_name: The name of the route calculator resource that you want to use to calculate the route.

departure_position: The start position for the route. Defined in WGS 84 format: [longitude, latitude]. For example, [-123.115, 49.285] If you specify a departure that's not located on a road, Amazon Location moves the position to the nearest road. If Esri is the provider for your route calculator, specifying a route that is longer than 400 km returns a 400 RoutesValidationException error. Valid Values: [-180 to 180,-90 to 90]

destination_position: The finish position for the route. Defined in WGS 84 format: [longitude, latitude]. For example, [-122.339, 47.615] If you specify a destination that's not located on a road, Amazon Location moves the position to the nearest road. Valid Values: [-180 to 180,-90 to 90]

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"CarModeOptions": Specifies route preferences when traveling by Car, such as avoiding routes that use ferries or tolls. Requirements: TravelMode must be specified as Car.

"DepartNow": Sets the time of departure as the current time. Uses the current time to calculate a route. Otherwise, the best time of day to travel with the best traffic conditions is used to calculate the route. Default Value: false Valid Values: false | true

"DepartureTime": Specifies the desired time of departure. Uses the given time to calculate the route. Otherwise, the best time of day to travel with the best traffic conditions is used to calculate the route. Setting a departure time in the past returns a 400 ValidationException error. In ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ. For example, 2020–07-2T12:15:20.000Z+01:00

"DistanceUnit": Set the unit system to specify the distance. Default Value: Kilometers

"IncludeLegGeometry": Set to include the geometry details in the result for each path between a pair of positions. Default Value: false Valid Values: false | true

"TravelMode": Specifies the mode of transport when calculating a route. Used in estimating the speed of travel and road compatibility. The TravelMode you specify determines how you specify route preferences: If traveling by Car use the CarModeOptions parameter. If traveling by Truck use the TruckModeOptions parameter. Default Value: Car

"TruckModeOptions": Specifies route preferences when traveling by Truck, such as avoiding routes that use ferries or tolls, and truck specifications to consider when choosing an optimal road. Requirements: TravelMode must be specified as Truck.

"WaypointPositions": Specifies an ordered list of up to 23 intermediate positions to include along a route between the departure position and destination position. For example, from the DeparturePosition [-123.115, 49.285], the route follows the order that the waypoint positions are given [[-122.757, 49.0021],[-122.349, 47.620]] If you specify a waypoint position that's not located on a road, Amazon Location moves the position to the nearest road. Specifying more than 23 waypoints returns a 400 ValidationException error. If Esri is the provider for your route calculator, specifying a route that is longer than 400 km returns a 400 RoutesValidationException error. Valid Values: [-180 to 180,-90 to 90]

Main.Location.create_geofence_collection — Method

create_geofence_collection(collection_name)
create_geofence_collection(collection_name, params::Dict{String,<:Any})

Creates a geofence collection, which manages and stores geofences.

Arguments

collection_name: A custom name for the geofence collection. Requirements: Contain only alphanumeric characters (A–Z, a–z, 0–9), hyphens (-), periods (.), and underscores (_). Must be a unique geofence collection name. No spaces allowed. For example, ExampleGeofenceCollection.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": An optional description for the geofence collection.

"KmsKeyId": A key identifier for an AWS KMS customer managed key. Enter a key ID, key ARN, alias name, or alias ARN.

"PricingPlan": Optionally specifies the pricing plan for the geofence collection. Defaults to RequestBasedUsage. For additional details and restrictions on each pricing plan option, see the Amazon Location Service pricing page.

"PricingPlanDataSource": Specifies the data provider for the geofence collection. Required value for the following pricing plans: MobileAssetTracking | MobileAssetManagement For more information about Data Providers, and Pricing plans, see the Amazon Location Service product page. Amazon Location Service only uses PricingPlanDataSource to calculate billing for your geofence collection. Your data won't be shared with the data provider, and will remain in your AWS account or Region unless you move it. Valid Values: Esri | Here

"Tags": Applies one or more tags to the geofence collection. A tag is a key-value pair helps manage, identify, search, and filter your resources by labelling them. Format: "key" : "value" Restrictions: Maximum 50 tags per resource Each resource tag must be unique with a maximum of one value. Maximum key length: 128 Unicode characters in UTF-8 Maximum value length: 256 Unicode characters in UTF-8 Can use alphanumeric characters (A–Z, a–z, 0–9), and the following characters: + - = . _ : / @. Cannot use "aws:" as a prefix for a key.

Main.Location.create_map — Method

create_map(configuration, map_name)
create_map(configuration, map_name, params::Dict{String,<:Any})

Creates a map resource in your AWS account, which provides map tiles of different styles sourced from global location data providers.

Arguments

configuration: Specifies the map style selected from an available data provider.

map_name: The name for the map resource. Requirements: Must contain only alphanumeric characters (A–Z, a–z, 0–9), hyphens (-), periods (.), and underscores (_). Must be a unique map resource name. No spaces allowed. For example, ExampleMap.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": An optional description for the map resource.

"PricingPlan": Optionally specifies the pricing plan for the map resource. Defaults to RequestBasedUsage. For additional details and restrictions on each pricing plan option, see Amazon Location Service pricing.

"Tags": Applies one or more tags to the map resource. A tag is a key-value pair helps manage, identify, search, and filter your resources by labelling them. Format: "key" : "value" Restrictions: Maximum 50 tags per resource Each resource tag must be unique with a maximum of one value. Maximum key length: 128 Unicode characters in UTF-8 Maximum value length: 256 Unicode characters in UTF-8 Can use alphanumeric characters (A–Z, a–z, 0–9), and the following characters: + - = . _ : / @. Cannot use "aws:" as a prefix for a key.

Main.Location.create_place_index — Method

create_place_index(data_source, index_name)
create_place_index(data_source, index_name, params::Dict{String,<:Any})

Creates a place index resource in your AWS account. Use a place index resource to geocode addresses and other text queries by using the SearchPlaceIndexForText operation, and reverse geocode coordinates by using the SearchPlaceIndexForPosition operation, and enable autosuggestions by using the SearchPlaceIndexForSuggestions operation.

Arguments

data_source: Specifies the geospatial data provider for the new place index. This field is case-sensitive. Enter the valid values as shown. For example, entering HERE returns an error. Valid values include: Esri – For additional information about Esri's coverage in your region of interest, see Esri details on geocoding coverage. Here – For additional information about HERE Technologies' coverage in your region of interest, see HERE details on goecoding coverage. If you specify HERE Technologies (Here) as the data provider, you may not store results for locations in Japan. For more information, see the AWS Service Terms for Amazon Location Service. For additional information , see Data providers on the Amazon Location Service Developer Guide.

index_name: The name of the place index resource. Requirements: Contain only alphanumeric characters (A–Z, a–z, 0–9), hyphens (-), periods (.), and underscores (_). Must be a unique place index resource name. No spaces allowed. For example, ExamplePlaceIndex.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DataSourceConfiguration": Specifies the data storage option requesting Places.

"Description": The optional description for the place index resource.

"PricingPlan": Optionally specifies the pricing plan for the place index resource. Defaults to RequestBasedUsage. For additional details and restrictions on each pricing plan option, see Amazon Location Service pricing.

"Tags": Applies one or more tags to the place index resource. A tag is a key-value pair that helps you manage, identify, search, and filter your resources. Format: "key" : "value" Restrictions: Maximum 50 tags per resource. Each tag key must be unique and must have exactly one associated value. Maximum key length: 128 Unicode characters in UTF-8. Maximum value length: 256 Unicode characters in UTF-8. Can use alphanumeric characters (A–Z, a–z, 0–9), and the following characters: + - = . _ : / @ Cannot use "aws:" as a prefix for a key.

Main.Location.create_route_calculator — Method

create_route_calculator(calculator_name, data_source)
create_route_calculator(calculator_name, data_source, params::Dict{String,<:Any})

Creates a route calculator resource in your AWS account. You can send requests to a route calculator resource to estimate travel time, distance, and get directions. A route calculator sources traffic and road network data from your chosen data provider.

Arguments

calculator_name: The name of the route calculator resource. Requirements: Can use alphanumeric characters (A–Z, a–z, 0–9) , hyphens (-), periods (.), and underscores (_). Must be a unique Route calculator resource name. No spaces allowed. For example, ExampleRouteCalculator.

data_source: Specifies the data provider of traffic and road network data. This field is case-sensitive. Enter the valid values as shown. For example, entering HERE returns an error. Route calculators that use Esri as a data source only calculate routes that are shorter than 400 km. Valid values include: Esri – For additional information about Esri's coverage in your region of interest, see Esri details on street networks and traffic coverage. Here – For additional information about HERE Technologies' coverage in your region of interest, see HERE car routing coverage and HERE truck routing coverage. For additional information , see Data providers on the Amazon Location Service Developer Guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": The optional description for the route calculator resource.

"PricingPlan": Optionally specifies the pricing plan for the route calculator resource. Defaults to RequestBasedUsage. For additional details and restrictions on each pricing plan option, see Amazon Location Service pricing.

"Tags": Applies one or more tags to the route calculator resource. A tag is a key-value pair helps manage, identify, search, and filter your resources by labelling them. For example: { "tag1" : "value1", "tag2" : "value2"} Format: "key" : "value" Restrictions: Maximum 50 tags per resource Each resource tag must be unique with a maximum of one value. Maximum key length: 128 Unicode characters in UTF-8 Maximum value length: 256 Unicode characters in UTF-8 Can use alphanumeric characters (A–Z, a–z, 0–9), and the following characters: + - = . _ : / @. Cannot use "aws:" as a prefix for a key.

Main.Location.create_tracker — Method

create_tracker(tracker_name)
create_tracker(tracker_name, params::Dict{String,<:Any})

Creates a tracker resource in your AWS account, which lets you retrieve current and historical location of devices.

Arguments

tracker_name: The name for the tracker resource. Requirements: Contain only alphanumeric characters (A-Z, a-z, 0-9) , hyphens (-), periods (.), and underscores (_). Must be a unique tracker resource name. No spaces allowed. For example, ExampleTracker.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": An optional description for the tracker resource.

"KmsKeyId": A key identifier for an AWS KMS customer managed key. Enter a key ID, key ARN, alias name, or alias ARN.

"PositionFiltering": Specifies the position filtering for the tracker resource. Valid values: TimeBased - Location updates are evaluated against linked geofence collections, but not every location update is stored. If your update frequency is more often than 30 seconds, only one update per 30 seconds is stored for each unique device ID. DistanceBased - If the device has moved less than 30 m (98.4 ft), location updates are ignored. Location updates within this area are neither evaluated against linked geofence collections, nor stored. This helps control costs by reducing the number of geofence evaluations and historical device positions to paginate through. Distance-based filtering can also reduce the effects of GPS noise when displaying device trajectories on a map. AccuracyBased - If the device has moved less than the measured accuracy, location updates are ignored. For example, if two consecutive updates from a device have a horizontal accuracy of 5 m and 10 m, the second update is ignored if the device has moved less than 15 m. Ignored location updates are neither evaluated against linked geofence collections, nor stored. This can reduce the effects of GPS noise when displaying device trajectories on a map, and can help control your costs by reducing the number of geofence evaluations. This field is optional. If not specified, the default value is TimeBased.

"PricingPlan": Optionally specifies the pricing plan for the tracker resource. Defaults to RequestBasedUsage. For additional details and restrictions on each pricing plan option, see Amazon Location Service pricing.

"PricingPlanDataSource": Specifies the data provider for the tracker resource. Required value for the following pricing plans: MobileAssetTracking | MobileAssetManagement For more information about Data Providers, and Pricing plans, see the Amazon Location Service product page. Amazon Location Service only uses PricingPlanDataSource to calculate billing for your tracker resource. Your data will not be shared with the data provider, and will remain in your AWS account or Region unless you move it. Valid values: Esri | Here

"Tags": Applies one or more tags to the tracker resource. A tag is a key-value pair helps manage, identify, search, and filter your resources by labelling them. Format: "key" : "value" Restrictions: Maximum 50 tags per resource Each resource tag must be unique with a maximum of one value. Maximum key length: 128 Unicode characters in UTF-8 Maximum value length: 256 Unicode characters in UTF-8 Can use alphanumeric characters (A–Z, a–z, 0–9), and the following characters: + - = . _ : / @. Cannot use "aws:" as a prefix for a key.

Main.Location.delete_geofence_collection — Method

delete_geofence_collection(collection_name)
delete_geofence_collection(collection_name, params::Dict{String,<:Any})

Deletes a geofence collection from your AWS account. This operation deletes the resource permanently. If the geofence collection is the target of a tracker resource, the devices will no longer be monitored.

Arguments

collection_name: The name of the geofence collection to be deleted.

Main.Location.delete_map — Method

delete_map(map_name)
delete_map(map_name, params::Dict{String,<:Any})

Deletes a map resource from your AWS account. This operation deletes the resource permanently. If the map is being used in an application, the map may not render.

Arguments

map_name: The name of the map resource to be deleted.

Main.Location.delete_place_index — Method

delete_place_index(index_name)
delete_place_index(index_name, params::Dict{String,<:Any})

Deletes a place index resource from your AWS account. This operation deletes the resource permanently.

Arguments

index_name: The name of the place index resource to be deleted.

Main.Location.delete_route_calculator — Method

delete_route_calculator(calculator_name)
delete_route_calculator(calculator_name, params::Dict{String,<:Any})

Deletes a route calculator resource from your AWS account. This operation deletes the resource permanently.

Arguments

calculator_name: The name of the route calculator resource to be deleted.

Main.Location.delete_tracker — Method

delete_tracker(tracker_name)
delete_tracker(tracker_name, params::Dict{String,<:Any})

Deletes a tracker resource from your AWS account. This operation deletes the resource permanently. If the tracker resource is in use, you may encounter an error. Make sure that the target resource isn't a dependency for your applications.

Arguments

tracker_name: The name of the tracker resource to be deleted.

Main.Location.describe_geofence_collection — Method

describe_geofence_collection(collection_name)
describe_geofence_collection(collection_name, params::Dict{String,<:Any})

Retrieves the geofence collection details.

Arguments

collection_name: The name of the geofence collection.

Main.Location.describe_map — Method

describe_map(map_name)
describe_map(map_name, params::Dict{String,<:Any})

Retrieves the map resource details.

Arguments

map_name: The name of the map resource.

Main.Location.describe_place_index — Method

describe_place_index(index_name)
describe_place_index(index_name, params::Dict{String,<:Any})

Retrieves the place index resource details.

Arguments

index_name: The name of the place index resource.

Main.Location.describe_route_calculator — Method

describe_route_calculator(calculator_name)
describe_route_calculator(calculator_name, params::Dict{String,<:Any})

Retrieves the route calculator resource details.

Arguments

calculator_name: The name of the route calculator resource.

Main.Location.describe_tracker — Method

describe_tracker(tracker_name)
describe_tracker(tracker_name, params::Dict{String,<:Any})

Retrieves the tracker resource details.

Arguments

tracker_name: The name of the tracker resource.

Main.Location.disassociate_tracker_consumer — Method

disassociate_tracker_consumer(consumer_arn, tracker_name)
disassociate_tracker_consumer(consumer_arn, tracker_name, params::Dict{String,<:Any})

Removes the association between a tracker resource and a geofence collection. Once you unlink a tracker resource from a geofence collection, the tracker positions will no longer be automatically evaluated against geofences.

Arguments

consumer_arn: The Amazon Resource Name (ARN) for the geofence collection to be disassociated from the tracker resource. Used when you need to specify a resource across all AWS. Format example: arn:aws:geo:region:account-id:geofence-collection/ExampleGeofenceCollectionConsumer

tracker_name: The name of the tracker resource to be dissociated from the consumer.

Main.Location.get_device_position — Method

get_device_position(device_id, tracker_name)
get_device_position(device_id, tracker_name, params::Dict{String,<:Any})

Retrieves a device's most recent position according to its sample time. Device positions are deleted after 30 days.

Arguments

device_id: The device whose position you want to retrieve.

tracker_name: The tracker resource receiving the position update.

Main.Location.get_device_position_history — Method

get_device_position_history(device_id, tracker_name)
get_device_position_history(device_id, tracker_name, params::Dict{String,<:Any})

Retrieves the device position history from a tracker resource within a specified range of time. Device positions are deleted after 30 days.

Arguments

device_id: The device whose position history you want to retrieve.

tracker_name: The tracker resource receiving the request for the device position history.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"EndTimeExclusive": Specify the end time for the position history in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ. By default, the value will be the time that the request is made. Requirement: The time specified for EndTimeExclusive must be after the time for StartTimeInclusive.

"NextToken": The pagination token specifying which page of results to return in the response. If no token is provided, the default page is the first page. Default value: null

"StartTimeInclusive": Specify the start time for the position history in ISO 8601 format: YYYY-MM-DDThh:mm:ss.sssZ. By default, the value will be 24 hours prior to the time that the request is made. Requirement: The time specified for StartTimeInclusive must be before EndTimeExclusive.

Main.Location.get_geofence — Method

get_geofence(collection_name, geofence_id)
get_geofence(collection_name, geofence_id, params::Dict{String,<:Any})

Retrieves the geofence details from a geofence collection.

Arguments

collection_name: The geofence collection storing the target geofence.

geofence_id: The geofence you're retrieving details for.

Main.Location.get_map_glyphs — Method

get_map_glyphs(font_stack, font_unicode_range, map_name)
get_map_glyphs(font_stack, font_unicode_range, map_name, params::Dict{String,<:Any})

Retrieves glyphs used to display labels on a map.

Arguments

font_stack: A comma-separated list of fonts to load glyphs from in order of preference. For example, Noto Sans Regular, Arial Unicode. Valid fonts stacks for Esri styles: VectorEsriDarkGrayCanvas – Ubuntu Medium Italic | Ubuntu Medium | Ubuntu Italic | Ubuntu Regular | Ubuntu Bold VectorEsriLightGrayCanvas – Ubuntu Italic | Ubuntu Regular | Ubuntu Light | Ubuntu Bold VectorEsriTopographic – Noto Sans Italic | Noto Sans Regular | Noto Sans Bold | Noto Serif Regular | Roboto Condensed Light Italic VectorEsriStreets – Arial Regular | Arial Italic | Arial Bold VectorEsriNavigation – Arial Regular | Arial Italic | Arial Bold Valid font stacks for HERE Technologies styles: VectorHereBerlin – Fira GO Regular | Fira GO Bold

font_unicode_range: A Unicode range of characters to download glyphs for. Each response will contain 256 characters. For example, 0–255 includes all characters from range U+0000 to 00FF. Must be aligned to multiples of 256.

map_name: The map resource associated with the glyph ﬁle.

Main.Location.get_map_sprites — Method

get_map_sprites(file_name, map_name)
get_map_sprites(file_name, map_name, params::Dict{String,<:Any})

Retrieves the sprite sheet corresponding to a map resource. The sprite sheet is a PNG image paired with a JSON document describing the offsets of individual icons that will be displayed on a rendered map.

Arguments

file_name: The name of the sprite ﬁle. Use the following ﬁle names for the sprite sheet: sprites.png sprites@2x.png for high pixel density displays For the JSON document contain image offsets. Use the following ﬁle names: sprites.json sprites@2x.json for high pixel density displays

map_name: The map resource associated with the sprite ﬁle.

Main.Location.get_map_style_descriptor — Method

get_map_style_descriptor(map_name)
get_map_style_descriptor(map_name, params::Dict{String,<:Any})

Retrieves the map style descriptor from a map resource. The style descriptor contains speciﬁcations on how features render on a map. For example, what data to display, what order to display the data in, and the style for the data. Style descriptors follow the Mapbox Style Specification.

Arguments

map_name: The map resource to retrieve the style descriptor from.

Main.Location.get_map_tile — Method

get_map_tile(map_name, x, y, z)
get_map_tile(map_name, x, y, z, params::Dict{String,<:Any})

Retrieves a vector data tile from the map resource. Map tiles are used by clients to render a map. they're addressed using a grid arrangement with an X coordinate, Y coordinate, and Z (zoom) level. The origin (0, 0) is the top left of the map. Increasing the zoom level by 1 doubles both the X and Y dimensions, so a tile containing data for the entire world at (0/0/0) will be split into 4 tiles at zoom 1 (1/0/0, 1/0/1, 1/1/0, 1/1/1).

Arguments

map_name: The map resource to retrieve the map tiles from.

x: The X axis value for the map tile.

y: The Y axis value for the map tile.

z: The zoom value for the map tile.

Main.Location.list_device_positions — Method

list_device_positions(tracker_name)
list_device_positions(tracker_name, params::Dict{String,<:Any})

A batch request to retrieve all device positions.

Arguments

tracker_name: The tracker resource containing the requested devices.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": An optional limit for the number of entries returned in a single call. Default value: 100

"NextToken": The pagination token specifying which page of results to return in the response. If no token is provided, the default page is the first page. Default value: null

Main.Location.list_geofence_collections — Method

list_geofence_collections()
list_geofence_collections(params::Dict{String,<:Any})

Lists geofence collections in your AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": An optional limit for the number of resources returned in a single call. Default value: 100

"NextToken": The pagination token specifying which page of results to return in the response. If no token is provided, the default page is the first page. Default value: null

Main.Location.list_geofences — Method

list_geofences(collection_name)
list_geofences(collection_name, params::Dict{String,<:Any})

Lists geofences stored in a given geofence collection.

Arguments

collection_name: The name of the geofence collection storing the list of geofences.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"NextToken": The pagination token specifying which page of results to return in the response. If no token is provided, the default page is the first page. Default value: null

Main.Location.list_maps — Method

list_maps()
list_maps(params::Dict{String,<:Any})

Lists map resources in your AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": An optional limit for the number of resources returned in a single call. Default value: 100

"NextToken": The pagination token specifying which page of results to return in the response. If no token is provided, the default page is the first page. Default value: null

Main.Location.list_place_indexes — Method

list_place_indexes()
list_place_indexes(params::Dict{String,<:Any})

Lists place index resources in your AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": An optional limit for the maximum number of results returned in a single call. Default value: 100

"NextToken": The pagination token specifying which page of results to return in the response. If no token is provided, the default page is the first page. Default value: null

Main.Location.list_route_calculators — Method

list_route_calculators()
list_route_calculators(params::Dict{String,<:Any})

Lists route calculator resources in your AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": An optional maximum number of results returned in a single call. Default Value: 100

"NextToken": The pagination token specifying which page of results to return in the response. If no token is provided, the default page is the first page. Default Value: null

Main.Location.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Returns a list of tags that are applied to the specified Amazon Location resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource whose tags you want to retrieve. Format example: arn:aws:geo:region:account-id:resourcetype/ExampleResource

Main.Location.list_tracker_consumers — Method

list_tracker_consumers(tracker_name)
list_tracker_consumers(tracker_name, params::Dict{String,<:Any})

Lists geofence collections currently associated to the given tracker resource.

Arguments

tracker_name: The tracker resource whose associated geofence collections you want to list.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": An optional limit for the number of resources returned in a single call. Default value: 100

"NextToken": The pagination token specifying which page of results to return in the response. If no token is provided, the default page is the first page. Default value: null

Main.Location.list_trackers — Method

list_trackers()
list_trackers(params::Dict{String,<:Any})

Lists tracker resources in your AWS account.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": An optional limit for the number of resources returned in a single call. Default value: 100

"NextToken": The pagination token specifying which page of results to return in the response. If no token is provided, the default page is the first page. Default value: null

Main.Location.put_geofence — Method

put_geofence(collection_name, geofence_id, geometry)
put_geofence(collection_name, geofence_id, geometry, params::Dict{String,<:Any})

Stores a geofence geometry in a given geofence collection, or updates the geometry of an existing geofence if a geofence ID is included in the request.

Arguments

collection_name: The geofence collection to store the geofence in.

geofence_id: An identifier for the geofence. For example, ExampleGeofence-1.

geometry: Contains the polygon details to specify the position of the geofence. Each geofence polygon can have a maximum of 1,000 vertices.

Main.Location.search_place_index_for_position — Method

search_place_index_for_position(index_name, position)
search_place_index_for_position(index_name, position, params::Dict{String,<:Any})

Reverse geocodes a given coordinate and returns a legible address. Allows you to search for Places or points of interest near a given position.

Arguments

index_name: The name of the place index resource you want to use for the search.

position: Specifies the longitude and latitude of the position to query. This parameter must contain a pair of numbers. The first number represents the X coordinate, or longitude; the second number represents the Y coordinate, or latitude. For example, [-123.1174, 49.2847] represents a position with longitude -123.1174 and latitude 49.2847.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Language": The preferred language used to return results. The value must be a valid BCP 47 language tag, for example, en for English. This setting affects the languages used in the results. It does not change which results are returned. If the language is not specified, or not supported for a particular result, the partner automatically chooses a language for the result.

"MaxResults": An optional parameter. The maximum number of results returned per request. Default value: 50

Main.Location.search_place_index_for_suggestions — Method

search_place_index_for_suggestions(index_name, text)
search_place_index_for_suggestions(index_name, text, params::Dict{String,<:Any})

Generates suggestions for addresses and points of interest based on partial or misspelled free-form text. This operation is also known as autocomplete, autosuggest, or fuzzy matching. Optional parameters let you narrow your search results by bounding box or country, or bias your search toward a specific position on the globe. You can search for suggested place names near a specified position by using BiasPosition, or filter results within a bounding box by using FilterBBox. These parameters are mutually exclusive; using both BiasPosition and FilterBBox in the same command returns an error.

Arguments

index_name: The name of the place index resource you want to use for the search.

text: The free-form partial text to use to generate place suggestions. For example, eiffel tow.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"BiasPosition": An optional parameter that indicates a preference for place suggestions that are closer to a specified position. If provided, this parameter must contain a pair of numbers. The first number represents the X coordinate, or longitude; the second number represents the Y coordinate, or latitude. For example, [-123.1174, 49.2847] represents the position with longitude -123.1174 and latitude 49.2847. BiasPosition and FilterBBox are mutually exclusive. Specifying both options results in an error.

"FilterBBox": An optional parameter that limits the search results by returning only suggestions within a specified bounding box. If provided, this parameter must contain a total of four consecutive numbers in two pairs. The first pair of numbers represents the X and Y coordinates (longitude and latitude, respectively) of the southwest corner of the bounding box; the second pair of numbers represents the X and Y coordinates (longitude and latitude, respectively) of the northeast corner of the bounding box. For example, [-12.7935, -37.4835, -12.0684, -36.9542] represents a bounding box where the southwest corner has longitude -12.7935 and latitude -37.4835, and the northeast corner has longitude -12.0684 and latitude -36.9542. FilterBBox and BiasPosition are mutually exclusive. Specifying both options results in an error.

"FilterCountries": An optional parameter that limits the search results by returning only suggestions within the provided list of countries. Use the ISO 3166 3-digit country code. For example, Australia uses three upper-case characters: AUS.

"Language": The preferred language used to return results. The value must be a valid BCP 47 language tag, for example, en for English. This setting affects the languages used in the results. It does not change which results are returned. If the language is not specified, or not supported for a particular result, the partner automatically chooses a language for the result. Used only when the partner selected is Here.

"MaxResults": An optional parameter. The maximum number of results returned per request. The default: 5

Main.Location.search_place_index_for_text — Method

search_place_index_for_text(index_name, text)
search_place_index_for_text(index_name, text, params::Dict{String,<:Any})

Geocodes free-form text, such as an address, name, city, or region to allow you to search for Places or points of interest. Optional parameters let you narrow your search results by bounding box or country, or bias your search toward a specific position on the globe. You can search for places near a given position using BiasPosition, or filter results within a bounding box using FilterBBox. Providing both parameters simultaneously returns an error. Search results are returned in order of highest to lowest relevance.

Arguments

index_name: The name of the place index resource you want to use for the search.

text: The address, name, city, or region to be used in the search in free-form text format. For example, 123 Any Street.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"BiasPosition": An optional parameter that indicates a preference for places that are closer to a specified position. If provided, this parameter must contain a pair of numbers. The first number represents the X coordinate, or longitude; the second number represents the Y coordinate, or latitude. For example, [-123.1174, 49.2847] represents the position with longitude -123.1174 and latitude 49.2847. BiasPosition and FilterBBox are mutually exclusive. Specifying both options results in an error.

"FilterBBox": An optional parameter that limits the search results by returning only places that are within the provided bounding box. If provided, this parameter must contain a total of four consecutive numbers in two pairs. The first pair of numbers represents the X and Y coordinates (longitude and latitude, respectively) of the southwest corner of the bounding box; the second pair of numbers represents the X and Y coordinates (longitude and latitude, respectively) of the northeast corner of the bounding box. For example, [-12.7935, -37.4835, -12.0684, -36.9542] represents a bounding box where the southwest corner has longitude -12.7935 and latitude -37.4835, and the northeast corner has longitude -12.0684 and latitude -36.9542. FilterBBox and BiasPosition are mutually exclusive. Specifying both options results in an error.

"FilterCountries": An optional parameter that limits the search results by returning only places that are in a specified list of countries. Valid values include ISO 3166 3-digit country codes. For example, Australia uses three upper-case characters: AUS.

"Language": The preferred language used to return results. The value must be a valid BCP 47 language tag, for example, en for English. This setting affects the languages used in the results. It does not change which results are returned. If the language is not specified, or not supported for a particular result, the partner automatically chooses a language for the result.

"MaxResults": An optional parameter. The maximum number of results returned per request. The default: 50

Main.Location.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Assigns one or more tags (key-value pairs) to the specified Amazon Location Service resource. &lt;p&gt;Tags can help you organize and categorize your resources. You can also use them to scope user permissions, by granting a user permission to access or change only resources with certain tag values.&lt;/p&gt; &lt;p&gt;You can use the &lt;code&gt;TagResource&lt;/code&gt; operation with an Amazon Location Service resource that already has tags. If you specify a new tag key for the resource, this tag is appended to the tags already associated with the resource. If you specify a tag key that's already associated with the resource, the new tag value that you specify replaces the previous value for that tag. &lt;/p&gt; &lt;p&gt;You can associate up to 50 tags with a resource.&lt;/p&gt;

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource whose tags you want to update. Format example: arn:aws:geo:region:account-id:resourcetype/ExampleResource

tags: Applies one or more tags to specific resource. A tag is a key-value pair that helps you manage, identify, search, and filter your resources. Format: "key" : "value" Restrictions: Maximum 50 tags per resource. Each tag key must be unique and must have exactly one associated value. Maximum key length: 128 Unicode characters in UTF-8. Maximum value length: 256 Unicode characters in UTF-8. Can use alphanumeric characters (A–Z, a–z, 0–9), and the following characters: + - = . _ : / @ Cannot use "aws:" as a prefix for a key.

Main.Location.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes one or more tags from the specified Amazon Location resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource from which you want to remove tags. Format example: arn:aws:geo:region:account-id:resourcetype/ExampleResource

tag_keys: The list of tag keys to remove from the specified resource.

Main.Location.update_geofence_collection — Method

update_geofence_collection(collection_name)
update_geofence_collection(collection_name, params::Dict{String,<:Any})

Updates the specified properties of a given geofence collection.

Arguments

collection_name: The name of the geofence collection to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": Updates the description for the geofence collection.

"PricingPlan": Updates the pricing plan for the geofence collection. For more information about each pricing plan option restrictions, see Amazon Location Service pricing.

"PricingPlanDataSource": Updates the data provider for the geofence collection. A required value for the following pricing plans: MobileAssetTracking| MobileAssetManagement For more information about data providers and pricing plans, see the Amazon Location Service product page. This can only be updated when updating the PricingPlan in the same request. Amazon Location Service uses PricingPlanDataSource to calculate billing for your geofence collection. Your data won't be shared with the data provider, and will remain in your AWS account and Region unless you move it.

Main.Location.update_map — Method

update_map(map_name)
update_map(map_name, params::Dict{String,<:Any})

Updates the specified properties of a given map resource.

Arguments

map_name: The name of the map resource to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": Updates the description for the map resource.

"PricingPlan": Updates the pricing plan for the map resource. For more information about each pricing plan option restrictions, see Amazon Location Service pricing.

Main.Location.update_place_index — Method

update_place_index(index_name)
update_place_index(index_name, params::Dict{String,<:Any})

Updates the specified properties of a given place index resource.

Arguments

index_name: The name of the place index resource to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"DataSourceConfiguration": Updates the data storage option for the place index resource.

"Description": Updates the description for the place index resource.

"PricingPlan": Updates the pricing plan for the place index resource. For more information about each pricing plan option restrictions, see Amazon Location Service pricing.

Main.Location.update_route_calculator — Method

update_route_calculator(calculator_name)
update_route_calculator(calculator_name, params::Dict{String,<:Any})

Updates the specified properties for a given route calculator resource.

Arguments

calculator_name: The name of the route calculator resource to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": Updates the description for the route calculator resource.

"PricingPlan": Updates the pricing plan for the route calculator resource. For more information about each pricing plan option restrictions, see Amazon Location Service pricing.

Main.Location.update_tracker — Method

update_tracker(tracker_name)
update_tracker(tracker_name, params::Dict{String,<:Any})

Updates the specified properties of a given tracker resource.

Arguments

tracker_name: The name of the tracker resource to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Description": Updates the description for the tracker resource.

"PositionFiltering": Updates the position filtering for the tracker resource. Valid values: TimeBased - Location updates are evaluated against linked geofence collections, but not every location update is stored. If your update frequency is more often than 30 seconds, only one update per 30 seconds is stored for each unique device ID. DistanceBased - If the device has moved less than 30 m (98.4 ft), location updates are ignored. Location updates within this distance are neither evaluated against linked geofence collections, nor stored. This helps control costs by reducing the number of geofence evaluations and historical device positions to paginate through. Distance-based filtering can also reduce the effects of GPS noise when displaying device trajectories on a map. AccuracyBased - If the device has moved less than the measured accuracy, location updates are ignored. For example, if two consecutive updates from a device have a horizontal accuracy of 5 m and 10 m, the second update is ignored if the device has moved less than 15 m. Ignored location updates are neither evaluated against linked geofence collections, nor stored. This helps educe the effects of GPS noise when displaying device trajectories on a map, and can help control costs by reducing the number of geofence evaluations.

"PricingPlan": Updates the pricing plan for the tracker resource. For more information about each pricing plan option restrictions, see Amazon Location Service pricing.

"PricingPlanDataSource": Updates the data provider for the tracker resource. A required value for the following pricing plans: MobileAssetTracking| MobileAssetManagement For more information about data providers and pricing plans, see the Amazon Location Service product page This can only be updated when updating the PricingPlan in the same request. Amazon Location Service uses PricingPlanDataSource to calculate billing for your tracker resource. Your data won't be shared with the data provider, and will remain in your AWS account and Region unless you move it.