Medialive

This page documents function available when using the Medialive module, created with @service Medialive.

Index

Main.Medialive.accept_input_device_transfer

Main.Medialive.batch_delete

Main.Medialive.batch_start

Main.Medialive.batch_stop

Main.Medialive.batch_update_schedule

Main.Medialive.cancel_input_device_transfer

Main.Medialive.claim_device

Main.Medialive.create_channel

Main.Medialive.create_cloud_watch_alarm_template

Main.Medialive.create_cloud_watch_alarm_template_group

Main.Medialive.create_event_bridge_rule_template

Main.Medialive.create_event_bridge_rule_template_group

Main.Medialive.create_input

Main.Medialive.create_input_security_group

Main.Medialive.create_multiplex

Main.Medialive.create_multiplex_program

Main.Medialive.create_partner_input

Main.Medialive.create_signal_map

Main.Medialive.create_tags

Main.Medialive.delete_channel

Main.Medialive.delete_cloud_watch_alarm_template

Main.Medialive.delete_cloud_watch_alarm_template_group

Main.Medialive.delete_event_bridge_rule_template

Main.Medialive.delete_event_bridge_rule_template_group

Main.Medialive.delete_input

Main.Medialive.delete_input_security_group

Main.Medialive.delete_multiplex

Main.Medialive.delete_multiplex_program

Main.Medialive.delete_reservation

Main.Medialive.delete_schedule

Main.Medialive.delete_signal_map

Main.Medialive.delete_tags

Main.Medialive.describe_account_configuration

Main.Medialive.describe_channel

Main.Medialive.describe_input

Main.Medialive.describe_input_device

Main.Medialive.describe_input_device_thumbnail

Main.Medialive.describe_input_security_group

Main.Medialive.describe_multiplex

Main.Medialive.describe_multiplex_program

Main.Medialive.describe_offering

Main.Medialive.describe_reservation

Main.Medialive.describe_schedule

Main.Medialive.describe_thumbnails

Main.Medialive.get_cloud_watch_alarm_template

Main.Medialive.get_cloud_watch_alarm_template_group

Main.Medialive.get_event_bridge_rule_template

Main.Medialive.get_event_bridge_rule_template_group

Main.Medialive.get_signal_map

Main.Medialive.list_channels

Main.Medialive.list_cloud_watch_alarm_template_groups

Main.Medialive.list_cloud_watch_alarm_templates

Main.Medialive.list_event_bridge_rule_template_groups

Main.Medialive.list_event_bridge_rule_templates

Main.Medialive.list_input_device_transfers

Main.Medialive.list_input_devices

Main.Medialive.list_input_security_groups

Main.Medialive.list_inputs

Main.Medialive.list_multiplex_programs

Main.Medialive.list_multiplexes

Main.Medialive.list_offerings

Main.Medialive.list_reservations

Main.Medialive.list_signal_maps

Main.Medialive.list_tags_for_resource

Main.Medialive.purchase_offering

Main.Medialive.reboot_input_device

Main.Medialive.reject_input_device_transfer

Main.Medialive.restart_channel_pipelines

Main.Medialive.start_channel

Main.Medialive.start_delete_monitor_deployment

Main.Medialive.start_input_device

Main.Medialive.start_input_device_maintenance_window

Main.Medialive.start_monitor_deployment

Main.Medialive.start_multiplex

Main.Medialive.start_update_signal_map

Main.Medialive.stop_channel

Main.Medialive.stop_input_device

Main.Medialive.stop_multiplex

Main.Medialive.transfer_input_device

Main.Medialive.update_account_configuration

Main.Medialive.update_channel

Main.Medialive.update_channel_class

Main.Medialive.update_cloud_watch_alarm_template

Main.Medialive.update_cloud_watch_alarm_template_group

Main.Medialive.update_event_bridge_rule_template

Main.Medialive.update_event_bridge_rule_template_group

Main.Medialive.update_input

Main.Medialive.update_input_device

Main.Medialive.update_input_security_group

Main.Medialive.update_multiplex

Main.Medialive.update_multiplex_program

Main.Medialive.update_reservation

Documentation

Main.Medialive.accept_input_device_transfer — Method

accept_input_device_transfer(input_device_id)
accept_input_device_transfer(input_device_id, params::Dict{String,<:Any})

Accept an incoming input device transfer. The ownership of the device will transfer to your AWS account.

Arguments

input_device_id: The unique ID of the input device to accept. For example, hd-123456789abcdef.

Main.Medialive.batch_delete — Method

batch_delete()
batch_delete(params::Dict{String,<:Any})

Starts delete of resources.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"channelIds": List of channel IDs

"inputIds": List of input IDs

"inputSecurityGroupIds": List of input security group IDs

"multiplexIds": List of multiplex IDs

Main.Medialive.batch_start — Method

batch_start()
batch_start(params::Dict{String,<:Any})

Starts existing resources

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"channelIds": List of channel IDs

"multiplexIds": List of multiplex IDs

Main.Medialive.batch_stop — Method

batch_stop()
batch_stop(params::Dict{String,<:Any})

Stops running resources

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"channelIds": List of channel IDs

"multiplexIds": List of multiplex IDs

Main.Medialive.batch_update_schedule — Method

batch_update_schedule(channel_id)
batch_update_schedule(channel_id, params::Dict{String,<:Any})

Update a channel schedule

Arguments

channel_id: Id of the channel whose schedule is being updated.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"creates": Schedule actions to create in the schedule.

"deletes": Schedule actions to delete from the schedule.

Main.Medialive.cancel_input_device_transfer — Method

cancel_input_device_transfer(input_device_id)
cancel_input_device_transfer(input_device_id, params::Dict{String,<:Any})

Cancel an input device transfer that you have requested.

Arguments

input_device_id: The unique ID of the input device to cancel. For example, hd-123456789abcdef.

Main.Medialive.claim_device — Method

claim_device()
claim_device(params::Dict{String,<:Any})

Send a request to claim an AWS Elemental device that you have purchased from a third-party vendor. After the request succeeds, you will own the device.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"id": The id of the device you want to claim.

Main.Medialive.create_channel — Method

create_channel()
create_channel(params::Dict{String,<:Any})

Creates a new channel

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"cdiInputSpecification": Specification of CDI inputs for this channel

"channelClass": The class for this channel. STANDARD for a channel with two pipelines or SINGLE_PIPELINE for a channel with one pipeline.

"destinations":

"encoderSettings":

"inputAttachments": List of input attachments for channel.

"inputSpecification": Specification of network and file inputs for this channel

"logLevel": The log level to write to CloudWatch Logs.

"maintenance": Maintenance settings for this channel.

"name": Name of channel.

"requestId": Unique request ID to be specified. This is needed to prevent retries from

creating multiple resources.

"reserved": Deprecated field that's only usable by whitelisted customers.

"roleArn": An optional Amazon Resource Name (ARN) of the role to assume when running the Channel.

"tags": A collection of key-value pairs.

"vpc": Settings for the VPC outputs

Main.Medialive.create_cloud_watch_alarm_template — Method

create_cloud_watch_alarm_template(comparison_operator, evaluation_periods, group_identifier, metric_name, name, period, statistic, target_resource_type, threshold, treat_missing_data)
create_cloud_watch_alarm_template(comparison_operator, evaluation_periods, group_identifier, metric_name, name, period, statistic, target_resource_type, threshold, treat_missing_data, params::Dict{String,<:Any})

Creates a cloudwatch alarm template to dynamically generate cloudwatch metric alarms on targeted resource types.

Arguments

comparison_operator:

evaluation_periods: The number of periods over which data is compared to the specified threshold.

group_identifier: A cloudwatch alarm template group's identifier. Can be either be its id or current name.

metric_name: The name of the metric associated with the alarm. Must be compatible with targetResourceType.

name: A resource's name. Names must be unique within the scope of a resource type in a specific region.

period: The period, in seconds, over which the specified statistic is applied.

statistic:

target_resource_type:

threshold: The threshold value to compare with the specified statistic.

treat_missing_data:

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"datapointsToAlarm": The number of datapoints within the evaluation period that must be breaching to trigger the alarm.

"description": A resource's optional description.

"tags":

Main.Medialive.create_cloud_watch_alarm_template_group — Method

create_cloud_watch_alarm_template_group(name)
create_cloud_watch_alarm_template_group(name, params::Dict{String,<:Any})

Creates a cloudwatch alarm template group to group your cloudwatch alarm templates and to attach to signal maps for dynamically creating alarms.

Arguments

name: A resource's name. Names must be unique within the scope of a resource type in a specific region.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A resource's optional description.

"tags":

Main.Medialive.create_event_bridge_rule_template — Method

create_event_bridge_rule_template(event_type, group_identifier, name)
create_event_bridge_rule_template(event_type, group_identifier, name, params::Dict{String,<:Any})

Creates an eventbridge rule template to monitor events and send notifications to your targeted resources.

Arguments

event_type:

group_identifier: An eventbridge rule template group's identifier. Can be either be its id or current name.

name: A resource's name. Names must be unique within the scope of a resource type in a specific region.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A resource's optional description.

"eventTargets":

"tags":

Main.Medialive.create_event_bridge_rule_template_group — Method

create_event_bridge_rule_template_group(name)
create_event_bridge_rule_template_group(name, params::Dict{String,<:Any})

Creates an eventbridge rule template group to group your eventbridge rule templates and to attach to signal maps for dynamically creating notification rules.

Arguments

name: A resource's name. Names must be unique within the scope of a resource type in a specific region.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A resource's optional description.

"tags":

Main.Medialive.create_input — Method

create_input()
create_input(params::Dict{String,<:Any})

Create an input

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"destinations": Destination settings for PUSH type inputs.

"inputDevices": Settings for the devices.

"inputSecurityGroups": A list of security groups referenced by IDs to attach to the input.

"mediaConnectFlows": A list of the MediaConnect Flows that you want to use in this input. You can specify as few as one

Flow and presently, as many as two. The only requirement is when you have more than one is that each Flow is in a separate Availability Zone as this ensures your EML input is redundant to AZ issues.

"name": Name of the input.

"requestId": Unique identifier of the request to ensure the request is handled

exactly once in case of retries.

"roleArn": The Amazon Resource Name (ARN) of the role this input assumes during and after creation.

"sources": The source URLs for a PULL-type input. Every PULL type input needs

exactly two source URLs for redundancy. Only specify sources for PULL type Inputs. Leave Destinations empty.

"tags": A collection of key-value pairs.

"type":

"vpc":

Main.Medialive.create_input_security_group — Method

create_input_security_group()
create_input_security_group(params::Dict{String,<:Any})

Creates a Input Security Group

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"tags": A collection of key-value pairs.

"whitelistRules": List of IPv4 CIDR addresses to whitelist

Main.Medialive.create_multiplex — Method

create_multiplex(availability_zones, multiplex_settings, name, request_id)
create_multiplex(availability_zones, multiplex_settings, name, request_id, params::Dict{String,<:Any})

Create a new multiplex.

Arguments

availability_zones: A list of availability zones for the multiplex. You must specify exactly two.

multiplex_settings: Configuration for a multiplex event.

name: Name of multiplex.

request_id: Unique request ID. This prevents retries from creating multiple

resources.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"tags": A collection of key-value pairs.

Main.Medialive.create_multiplex_program — Method

create_multiplex_program(multiplex_id, multiplex_program_settings, program_name, request_id)
create_multiplex_program(multiplex_id, multiplex_program_settings, program_name, request_id, params::Dict{String,<:Any})

Create a new program in the multiplex.

Arguments

multiplex_id: ID of the multiplex where the program is to be created.

multiplex_program_settings: The settings for this multiplex program.

program_name: Name of multiplex program.

request_id: Unique request ID. This prevents retries from creating multiple

resources.

Main.Medialive.create_partner_input — Method

create_partner_input(input_id)
create_partner_input(input_id, params::Dict{String,<:Any})

Create a partner input

Arguments

input_id: Unique ID of the input.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"requestId": Unique identifier of the request to ensure the request is handled

exactly once in case of retries.

"tags": A collection of key-value pairs.

Main.Medialive.create_signal_map — Method

create_signal_map(discovery_entry_point_arn, name)
create_signal_map(discovery_entry_point_arn, name, params::Dict{String,<:Any})

Initiates the creation of a new signal map. Will discover a new mediaResourceMap based on the provided discoveryEntryPointArn.

Arguments

discovery_entry_point_arn: A top-level supported AWS resource ARN to discovery a signal map from.

name: A resource's name. Names must be unique within the scope of a resource type in a specific region.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"cloudWatchAlarmTemplateGroupIdentifiers":

"description": A resource's optional description.

"eventBridgeRuleTemplateGroupIdentifiers":

"tags":

Main.Medialive.create_tags — Method

create_tags(resource-arn)
create_tags(resource-arn, params::Dict{String,<:Any})

Create tags for a resource

Arguments

resource-arn:

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"tags":

Main.Medialive.delete_channel — Method

delete_channel(channel_id)
delete_channel(channel_id, params::Dict{String,<:Any})

Starts deletion of channel. The associated outputs are also deleted.

Arguments

channel_id: Unique ID of the channel.

Main.Medialive.delete_cloud_watch_alarm_template — Method

delete_cloud_watch_alarm_template(identifier)
delete_cloud_watch_alarm_template(identifier, params::Dict{String,<:Any})

Deletes a cloudwatch alarm template.

Arguments

identifier: A cloudwatch alarm template's identifier. Can be either be its id or current name.

Main.Medialive.delete_cloud_watch_alarm_template_group — Method

delete_cloud_watch_alarm_template_group(identifier)
delete_cloud_watch_alarm_template_group(identifier, params::Dict{String,<:Any})

Deletes a cloudwatch alarm template group. You must detach this group from all signal maps and ensure its existing templates are moved to another group or deleted.

Arguments

identifier: A cloudwatch alarm template group's identifier. Can be either be its id or current name.

Main.Medialive.delete_event_bridge_rule_template — Method

delete_event_bridge_rule_template(identifier)
delete_event_bridge_rule_template(identifier, params::Dict{String,<:Any})

Deletes an eventbridge rule template.

Arguments

identifier: An eventbridge rule template's identifier. Can be either be its id or current name.

Main.Medialive.delete_event_bridge_rule_template_group — Method

delete_event_bridge_rule_template_group(identifier)
delete_event_bridge_rule_template_group(identifier, params::Dict{String,<:Any})

Deletes an eventbridge rule template group. You must detach this group from all signal maps and ensure its existing templates are moved to another group or deleted.

Arguments

identifier: An eventbridge rule template group's identifier. Can be either be its id or current name.

Main.Medialive.delete_input — Method

delete_input(input_id)
delete_input(input_id, params::Dict{String,<:Any})

Deletes the input end point

Arguments

input_id: Unique ID of the input

Main.Medialive.delete_input_security_group — Method

delete_input_security_group(input_security_group_id)
delete_input_security_group(input_security_group_id, params::Dict{String,<:Any})

Deletes an Input Security Group

Arguments

input_security_group_id: The Input Security Group to delete

Main.Medialive.delete_multiplex — Method

delete_multiplex(multiplex_id)
delete_multiplex(multiplex_id, params::Dict{String,<:Any})

Delete a multiplex. The multiplex must be idle.

Arguments

multiplex_id: The ID of the multiplex.

Main.Medialive.delete_multiplex_program — Method

delete_multiplex_program(multiplex_id, program_name)
delete_multiplex_program(multiplex_id, program_name, params::Dict{String,<:Any})

Delete a program from a multiplex.

Arguments

multiplex_id: The ID of the multiplex that the program belongs to.

program_name: The multiplex program name.

Main.Medialive.delete_reservation — Method

delete_reservation(reservation_id)
delete_reservation(reservation_id, params::Dict{String,<:Any})

Delete an expired reservation.

Arguments

reservation_id: Unique reservation ID, e.g. '1234567'

Main.Medialive.delete_schedule — Method

delete_schedule(channel_id)
delete_schedule(channel_id, params::Dict{String,<:Any})

Delete all schedule actions on a channel.

Arguments

channel_id: Id of the channel whose schedule is being deleted.

Main.Medialive.delete_signal_map — Method

delete_signal_map(identifier)
delete_signal_map(identifier, params::Dict{String,<:Any})

Deletes the specified signal map.

Arguments

identifier: A signal map's identifier. Can be either be its id or current name.

Main.Medialive.delete_tags — Method

delete_tags(resource-arn, tag_keys)
delete_tags(resource-arn, tag_keys, params::Dict{String,<:Any})

Removes tags for a resource

Arguments

resource-arn:

tag_keys: An array of tag keys to delete

Main.Medialive.describe_account_configuration — Method

describe_account_configuration()
describe_account_configuration(params::Dict{String,<:Any})

Describe account configuration

Main.Medialive.describe_channel — Method

describe_channel(channel_id)
describe_channel(channel_id, params::Dict{String,<:Any})

Gets details about a channel

Arguments

channel_id: channel ID

Main.Medialive.describe_input — Method

describe_input(input_id)
describe_input(input_id, params::Dict{String,<:Any})

Produces details about an input

Arguments

input_id: Unique ID of the input

Main.Medialive.describe_input_device — Method

describe_input_device(input_device_id)
describe_input_device(input_device_id, params::Dict{String,<:Any})

Gets the details for the input device

Arguments

input_device_id: The unique ID of this input device. For example, hd-123456789abcdef.

Main.Medialive.describe_input_device_thumbnail — Method

describe_input_device_thumbnail(accept, input_device_id)
describe_input_device_thumbnail(accept, input_device_id, params::Dict{String,<:Any})

Get the latest thumbnail data for the input device.

Arguments

accept: The HTTP Accept header. Indicates the requested type for the thumbnail.

input_device_id: The unique ID of this input device. For example, hd-123456789abcdef.

Main.Medialive.describe_input_security_group — Method

describe_input_security_group(input_security_group_id)
describe_input_security_group(input_security_group_id, params::Dict{String,<:Any})

Produces a summary of an Input Security Group

Arguments

input_security_group_id: The id of the Input Security Group to describe

Main.Medialive.describe_multiplex — Method

describe_multiplex(multiplex_id)
describe_multiplex(multiplex_id, params::Dict{String,<:Any})

Gets details about a multiplex.

Arguments

multiplex_id: The ID of the multiplex.

Main.Medialive.describe_multiplex_program — Method

describe_multiplex_program(multiplex_id, program_name)
describe_multiplex_program(multiplex_id, program_name, params::Dict{String,<:Any})

Get the details for a program in a multiplex.

Arguments

multiplex_id: The ID of the multiplex that the program belongs to.

program_name: The name of the program.

Main.Medialive.describe_offering — Method

describe_offering(offering_id)
describe_offering(offering_id, params::Dict{String,<:Any})

Get details for an offering.

Arguments

offering_id: Unique offering ID, e.g. '87654321'

Main.Medialive.describe_reservation — Method

describe_reservation(reservation_id)
describe_reservation(reservation_id, params::Dict{String,<:Any})

Get details for a reservation.

Arguments

reservation_id: Unique reservation ID, e.g. '1234567'

Main.Medialive.describe_schedule — Method

describe_schedule(channel_id)
describe_schedule(channel_id, params::Dict{String,<:Any})

Get a channel schedule

Arguments

channel_id: Id of the channel whose schedule is being updated.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults":

"nextToken":

Main.Medialive.describe_thumbnails — Method

describe_thumbnails(channel_id, pipeline_id, thumbnail_type)
describe_thumbnails(channel_id, pipeline_id, thumbnail_type, params::Dict{String,<:Any})

Describe the latest thumbnails data.

Arguments

channel_id: Unique ID of the channel

pipeline_id: Pipeline ID ("0" or "1")

thumbnail_type: thumbnail type

Main.Medialive.get_cloud_watch_alarm_template — Method

get_cloud_watch_alarm_template(identifier)
get_cloud_watch_alarm_template(identifier, params::Dict{String,<:Any})

Retrieves the specified cloudwatch alarm template.

Arguments

identifier: A cloudwatch alarm template's identifier. Can be either be its id or current name.

Main.Medialive.get_cloud_watch_alarm_template_group — Method

get_cloud_watch_alarm_template_group(identifier)
get_cloud_watch_alarm_template_group(identifier, params::Dict{String,<:Any})

Retrieves the specified cloudwatch alarm template group.

Arguments

identifier: A cloudwatch alarm template group's identifier. Can be either be its id or current name.

Main.Medialive.get_event_bridge_rule_template — Method

get_event_bridge_rule_template(identifier)
get_event_bridge_rule_template(identifier, params::Dict{String,<:Any})

Retrieves the specified eventbridge rule template.

Arguments

identifier: An eventbridge rule template's identifier. Can be either be its id or current name.

Main.Medialive.get_event_bridge_rule_template_group — Method

get_event_bridge_rule_template_group(identifier)
get_event_bridge_rule_template_group(identifier, params::Dict{String,<:Any})

Retrieves the specified eventbridge rule template group.

Arguments

identifier: An eventbridge rule template group's identifier. Can be either be its id or current name.

Main.Medialive.get_signal_map — Method

get_signal_map(identifier)
get_signal_map(identifier, params::Dict{String,<:Any})

Retrieves the specified signal map.

Arguments

identifier: A signal map's identifier. Can be either be its id or current name.

Main.Medialive.list_channels — Method

list_channels()
list_channels(params::Dict{String,<:Any})

Produces list of channels that have been created

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults":

"nextToken":

Main.Medialive.list_cloud_watch_alarm_template_groups — Method

list_cloud_watch_alarm_template_groups()
list_cloud_watch_alarm_template_groups(params::Dict{String,<:Any})

Lists cloudwatch alarm template groups.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults":

"nextToken": A token used to retrieve the next set of results in paginated list responses.

"scope": Represents the scope of a resource, with options for all scopes, AWS provided resources, or local resources.

"signalMapIdentifier": A signal map's identifier. Can be either be its id or current name.

Main.Medialive.list_cloud_watch_alarm_templates — Method

list_cloud_watch_alarm_templates()
list_cloud_watch_alarm_templates(params::Dict{String,<:Any})

Lists cloudwatch alarm templates.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"groupIdentifier": A cloudwatch alarm template group's identifier. Can be either be its id or current name.

"maxResults":

"nextToken": A token used to retrieve the next set of results in paginated list responses.

"scope": Represents the scope of a resource, with options for all scopes, AWS provided resources, or local resources.

"signalMapIdentifier": A signal map's identifier. Can be either be its id or current name.

Main.Medialive.list_event_bridge_rule_template_groups — Method

list_event_bridge_rule_template_groups()
list_event_bridge_rule_template_groups(params::Dict{String,<:Any})

Lists eventbridge rule template groups.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults":

"nextToken": A token used to retrieve the next set of results in paginated list responses.

"signalMapIdentifier": A signal map's identifier. Can be either be its id or current name.

Main.Medialive.list_event_bridge_rule_templates — Method

list_event_bridge_rule_templates()
list_event_bridge_rule_templates(params::Dict{String,<:Any})

Lists eventbridge rule templates.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"groupIdentifier": An eventbridge rule template group's identifier. Can be either be its id or current name.

"maxResults":

"nextToken": A token used to retrieve the next set of results in paginated list responses.

"signalMapIdentifier": A signal map's identifier. Can be either be its id or current name.

Main.Medialive.list_input_device_transfers — Method

list_input_device_transfers(transfer_type)
list_input_device_transfers(transfer_type, params::Dict{String,<:Any})

List input devices that are currently being transferred. List input devices that you are transferring from your AWS account or input devices that another AWS account is transferring to you.

Arguments

transfer_type:

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults":

"nextToken":

Main.Medialive.list_input_devices — Method

list_input_devices()
list_input_devices(params::Dict{String,<:Any})

List input devices

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults":

"nextToken":

Main.Medialive.list_input_security_groups — Method

list_input_security_groups()
list_input_security_groups(params::Dict{String,<:Any})

Produces a list of Input Security Groups for an account

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults":

"nextToken":

Main.Medialive.list_inputs — Method

list_inputs()
list_inputs(params::Dict{String,<:Any})

Produces list of inputs that have been created

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults":

"nextToken":

Main.Medialive.list_multiplex_programs — Method

list_multiplex_programs(multiplex_id)
list_multiplex_programs(multiplex_id, params::Dict{String,<:Any})

List the programs that currently exist for a specific multiplex.

Arguments

multiplex_id: The ID of the multiplex that the programs belong to.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of items to return.

"nextToken": The token to retrieve the next page of results.

Main.Medialive.list_multiplexes — Method

list_multiplexes()
list_multiplexes(params::Dict{String,<:Any})

Retrieve a list of the existing multiplexes.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of items to return.

"nextToken": The token to retrieve the next page of results.

Main.Medialive.list_offerings — Method

list_offerings()
list_offerings(params::Dict{String,<:Any})

List offerings available for purchase.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"channelClass": Filter by channel class, 'STANDARD' or 'SINGLE_PIPELINE'

"channelConfiguration": Filter to offerings that match the configuration of an existing channel, e.g. '2345678' (a channel ID)

"codec": Filter by codec, 'AVC', 'HEVC', 'MPEG2', 'AUDIO', or 'LINK'

"duration": Filter by offering duration, e.g. '12'

"maxResults":

"maximumBitrate": Filter by bitrate, 'MAX10MBPS', 'MAX20MBPS', or 'MAX50MBPS'

"maximumFramerate": Filter by framerate, 'MAX30FPS' or 'MAX60FPS'

"nextToken":

"resolution": Filter by resolution, 'SD', 'HD', 'FHD', or 'UHD'

"resourceType": Filter by resource type, 'INPUT', 'OUTPUT', 'MULTIPLEX', or 'CHANNEL'

"specialFeature": Filter by special feature, 'ADVANCEDAUDIO' or 'AUDIONORMALIZATION'

"videoQuality": Filter by video quality, 'STANDARD', 'ENHANCED', or 'PREMIUM'

Main.Medialive.list_reservations — Method

list_reservations()
list_reservations(params::Dict{String,<:Any})

List purchased reservations.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"channelClass": Filter by channel class, 'STANDARD' or 'SINGLE_PIPELINE'

"codec": Filter by codec, 'AVC', 'HEVC', 'MPEG2', 'AUDIO', or 'LINK'

"maxResults":

"maximumBitrate": Filter by bitrate, 'MAX10MBPS', 'MAX20MBPS', or 'MAX50MBPS'

"maximumFramerate": Filter by framerate, 'MAX30FPS' or 'MAX60FPS'

"nextToken":

"resolution": Filter by resolution, 'SD', 'HD', 'FHD', or 'UHD'

"resourceType": Filter by resource type, 'INPUT', 'OUTPUT', 'MULTIPLEX', or 'CHANNEL'

"specialFeature": Filter by special feature, 'ADVANCEDAUDIO' or 'AUDIONORMALIZATION'

"videoQuality": Filter by video quality, 'STANDARD', 'ENHANCED', or 'PREMIUM'

Main.Medialive.list_signal_maps — Method

list_signal_maps()
list_signal_maps(params::Dict{String,<:Any})

Lists signal maps.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"cloudWatchAlarmTemplateGroupIdentifier": A cloudwatch alarm template group's identifier. Can be either be its id or current name.

"eventBridgeRuleTemplateGroupIdentifier": An eventbridge rule template group's identifier. Can be either be its id or current name.

"maxResults":

"nextToken": A token used to retrieve the next set of results in paginated list responses.

Main.Medialive.list_tags_for_resource — Method

list_tags_for_resource(resource-arn)
list_tags_for_resource(resource-arn, params::Dict{String,<:Any})

Produces list of tags that have been created for a resource

Arguments

resource-arn:

Main.Medialive.purchase_offering — Method

purchase_offering(count, offering_id)
purchase_offering(count, offering_id, params::Dict{String,<:Any})

Purchase an offering and create a reservation.

Arguments

count: Number of resources

offering_id: Offering to purchase, e.g. '87654321'

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"name": Name for the new reservation

"renewalSettings": Renewal settings for the reservation

"requestId": Unique request ID to be specified. This is needed to prevent retries from creating multiple resources.

"start": Requested reservation start time (UTC) in ISO-8601 format. The specified time must be between the first day of the current month and one year from now. If no value is given, the default is now.

"tags": A collection of key-value pairs

Main.Medialive.reboot_input_device — Method

reboot_input_device(input_device_id)
reboot_input_device(input_device_id, params::Dict{String,<:Any})

Send a reboot command to the specified input device. The device will begin rebooting within a few seconds of sending the command. When the reboot is complete, the device’s connection status will change to connected.

Arguments

input_device_id: The unique ID of the input device to reboot. For example, hd-123456789abcdef.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"force": Force a reboot of an input device. If the device is streaming, it will stop streaming and begin rebooting within a few seconds of sending the command. If the device was streaming prior to the reboot, the device will resume streaming when the reboot completes.

Main.Medialive.reject_input_device_transfer — Method

reject_input_device_transfer(input_device_id)
reject_input_device_transfer(input_device_id, params::Dict{String,<:Any})

Reject the transfer of the specified input device to your AWS account.

Arguments

input_device_id: The unique ID of the input device to reject. For example, hd-123456789abcdef.

Main.Medialive.restart_channel_pipelines — Method

restart_channel_pipelines(channel_id)
restart_channel_pipelines(channel_id, params::Dict{String,<:Any})

Restart pipelines in one channel that is currently running.

Arguments

channel_id: ID of channel

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"pipelineIds": An array of pipelines to restart in this channel. Format PIPELINE0 or PIPELINE1.

Main.Medialive.start_channel — Method

start_channel(channel_id)
start_channel(channel_id, params::Dict{String,<:Any})

Starts an existing channel

Arguments

channel_id: A request to start a channel

Main.Medialive.start_delete_monitor_deployment — Method

start_delete_monitor_deployment(identifier)
start_delete_monitor_deployment(identifier, params::Dict{String,<:Any})

Initiates a deployment to delete the monitor of the specified signal map.

Arguments

identifier: A signal map's identifier. Can be either be its id or current name.

Main.Medialive.start_input_device — Method

start_input_device(input_device_id)
start_input_device(input_device_id, params::Dict{String,<:Any})

Start an input device that is attached to a MediaConnect flow. (There is no need to start a device that is attached to a MediaLive input; MediaLive starts the device when the channel starts.)

Arguments

input_device_id: The unique ID of the input device to start. For example, hd-123456789abcdef.

Main.Medialive.start_input_device_maintenance_window — Method

start_input_device_maintenance_window(input_device_id)
start_input_device_maintenance_window(input_device_id, params::Dict{String,<:Any})

Start a maintenance window for the specified input device. Starting a maintenance window will give the device up to two hours to install software. If the device was streaming prior to the maintenance, it will resume streaming when the software is fully installed. Devices automatically install updates while they are powered on and their MediaLive channels are stopped. A maintenance window allows you to update a device without having to stop MediaLive channels that use the device. The device must remain powered on and connected to the internet for the duration of the maintenance.

Arguments

input_device_id: The unique ID of the input device to start a maintenance window for. For example, hd-123456789abcdef.

Main.Medialive.start_monitor_deployment — Method

start_monitor_deployment(identifier)
start_monitor_deployment(identifier, params::Dict{String,<:Any})

Initiates a deployment to deploy the latest monitor of the specified signal map.

Arguments

identifier: A signal map's identifier. Can be either be its id or current name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"dryRun":

Main.Medialive.start_multiplex — Method

start_multiplex(multiplex_id)
start_multiplex(multiplex_id, params::Dict{String,<:Any})

Start (run) the multiplex. Starting the multiplex does not start the channels. You must explicitly start each channel.

Arguments

multiplex_id: The ID of the multiplex.

Main.Medialive.start_update_signal_map — Method

start_update_signal_map(identifier)
start_update_signal_map(identifier, params::Dict{String,<:Any})

Initiates an update for the specified signal map. Will discover a new signal map if a changed discoveryEntryPointArn is provided.

Arguments

identifier: A signal map's identifier. Can be either be its id or current name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"cloudWatchAlarmTemplateGroupIdentifiers":

"description": A resource's optional description.

"discoveryEntryPointArn": A top-level supported AWS resource ARN to discovery a signal map from.

"eventBridgeRuleTemplateGroupIdentifiers":

"forceRediscovery": If true, will force a rediscovery of a signal map if an unchanged discoveryEntryPointArn is provided.

"name": A resource's name. Names must be unique within the scope of a resource type in a specific region.

Main.Medialive.stop_channel — Method

stop_channel(channel_id)
stop_channel(channel_id, params::Dict{String,<:Any})

Stops a running channel

Arguments

channel_id: A request to stop a running channel

Main.Medialive.stop_input_device — Method

stop_input_device(input_device_id)
stop_input_device(input_device_id, params::Dict{String,<:Any})

Stop an input device that is attached to a MediaConnect flow. (There is no need to stop a device that is attached to a MediaLive input; MediaLive automatically stops the device when the channel stops.)

Arguments

input_device_id: The unique ID of the input device to stop. For example, hd-123456789abcdef.

Main.Medialive.stop_multiplex — Method

stop_multiplex(multiplex_id)
stop_multiplex(multiplex_id, params::Dict{String,<:Any})

Stops a running multiplex. If the multiplex isn't running, this action has no effect.

Arguments

multiplex_id: The ID of the multiplex.

Main.Medialive.transfer_input_device — Method

transfer_input_device(input_device_id)
transfer_input_device(input_device_id, params::Dict{String,<:Any})

Start an input device transfer to another AWS account. After you make the request, the other account must accept or reject the transfer.

Arguments

input_device_id: The unique ID of this input device. For example, hd-123456789abcdef.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"targetCustomerId": The AWS account ID (12 digits) for the recipient of the device transfer.

"targetRegion": The target AWS region to transfer the device.

"transferMessage": An optional message for the recipient. Maximum 280 characters.

Main.Medialive.update_account_configuration — Method

update_account_configuration()
update_account_configuration(params::Dict{String,<:Any})

Update account configuration

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"accountConfiguration":

Main.Medialive.update_channel — Method

update_channel(channel_id)
update_channel(channel_id, params::Dict{String,<:Any})

Updates a channel.

Arguments

channel_id: channel ID

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"cdiInputSpecification": Specification of CDI inputs for this channel

"destinations": A list of output destinations for this channel.

"encoderSettings": The encoder settings for this channel.

"inputAttachments":

"inputSpecification": Specification of network and file inputs for this channel

"logLevel": The log level to write to CloudWatch Logs.

"maintenance": Maintenance settings for this channel.

"name": The name of the channel.

"roleArn": An optional Amazon Resource Name (ARN) of the role to assume when running the Channel. If you do not specify this on an update call but the role was previously set that role will be removed.

Main.Medialive.update_channel_class — Method

update_channel_class(channel_class, channel_id)
update_channel_class(channel_class, channel_id, params::Dict{String,<:Any})

Changes the class of the channel.

Arguments

channel_class: The channel class that you wish to update this channel to use.

channel_id: Channel Id of the channel whose class should be updated.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"destinations": A list of output destinations for this channel.

Main.Medialive.update_cloud_watch_alarm_template — Method

update_cloud_watch_alarm_template(identifier)
update_cloud_watch_alarm_template(identifier, params::Dict{String,<:Any})

Updates the specified cloudwatch alarm template.

Arguments

identifier: A cloudwatch alarm template's identifier. Can be either be its id or current name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"comparisonOperator":

"datapointsToAlarm": The number of datapoints within the evaluation period that must be breaching to trigger the alarm.

"description": A resource's optional description.

"evaluationPeriods": The number of periods over which data is compared to the specified threshold.

"groupIdentifier": A cloudwatch alarm template group's identifier. Can be either be its id or current name.

"metricName": The name of the metric associated with the alarm. Must be compatible with targetResourceType.

"name": A resource's name. Names must be unique within the scope of a resource type in a specific region.

"period": The period, in seconds, over which the specified statistic is applied.

"statistic":

"targetResourceType":

"threshold": The threshold value to compare with the specified statistic.

"treatMissingData":

Main.Medialive.update_cloud_watch_alarm_template_group — Method

update_cloud_watch_alarm_template_group(identifier)
update_cloud_watch_alarm_template_group(identifier, params::Dict{String,<:Any})

Updates the specified cloudwatch alarm template group.

Arguments

identifier: A cloudwatch alarm template group's identifier. Can be either be its id or current name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A resource's optional description.

Main.Medialive.update_event_bridge_rule_template — Method

update_event_bridge_rule_template(identifier)
update_event_bridge_rule_template(identifier, params::Dict{String,<:Any})

Updates the specified eventbridge rule template.

Arguments

identifier: An eventbridge rule template's identifier. Can be either be its id or current name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A resource's optional description.

"eventTargets":

"eventType":

"groupIdentifier": An eventbridge rule template group's identifier. Can be either be its id or current name.

"name": A resource's name. Names must be unique within the scope of a resource type in a specific region.

Main.Medialive.update_event_bridge_rule_template_group — Method

update_event_bridge_rule_template_group(identifier)
update_event_bridge_rule_template_group(identifier, params::Dict{String,<:Any})

Updates the specified eventbridge rule template group.

Arguments

identifier: An eventbridge rule template group's identifier. Can be either be its id or current name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A resource's optional description.

Main.Medialive.update_input — Method

update_input(input_id)
update_input(input_id, params::Dict{String,<:Any})

Updates an input.

Arguments

input_id: Unique ID of the input.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"destinations": Destination settings for PUSH type inputs.

"inputDevices": Settings for the devices.

"inputSecurityGroups": A list of security groups referenced by IDs to attach to the input.

"mediaConnectFlows": A list of the MediaConnect Flow ARNs that you want to use as the source of the input. You can specify as few as one

Flow and presently, as many as two. The only requirement is when you have more than one is that each Flow is in a separate Availability Zone as this ensures your EML input is redundant to AZ issues.

"name": Name of the input.

"roleArn": The Amazon Resource Name (ARN) of the role this input assumes during and after creation.

"sources": The source URLs for a PULL-type input. Every PULL type input needs

exactly two source URLs for redundancy. Only specify sources for PULL type Inputs. Leave Destinations empty.

Main.Medialive.update_input_device — Method

update_input_device(input_device_id)
update_input_device(input_device_id, params::Dict{String,<:Any})

Updates the parameters for the input device.

Arguments

input_device_id: The unique ID of the input device. For example, hd-123456789abcdef.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"availabilityZone": The Availability Zone you want associated with this input device.

"hdDeviceSettings": The settings that you want to apply to the HD input device.

"name": The name that you assigned to this input device (not the unique ID).

"uhdDeviceSettings": The settings that you want to apply to the UHD input device.

Main.Medialive.update_input_security_group — Method

update_input_security_group(input_security_group_id)
update_input_security_group(input_security_group_id, params::Dict{String,<:Any})

Update an Input Security Group's Whilelists.

Arguments

input_security_group_id: The id of the Input Security Group to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"tags": A collection of key-value pairs.

"whitelistRules": List of IPv4 CIDR addresses to whitelist

Main.Medialive.update_multiplex — Method

update_multiplex(multiplex_id)
update_multiplex(multiplex_id, params::Dict{String,<:Any})

Updates a multiplex.

Arguments

multiplex_id: ID of the multiplex to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"multiplexSettings": The new settings for a multiplex.

"name": Name of the multiplex.

Main.Medialive.update_multiplex_program — Method

update_multiplex_program(multiplex_id, program_name)
update_multiplex_program(multiplex_id, program_name, params::Dict{String,<:Any})

Update a program in a multiplex.

Arguments

multiplex_id: The ID of the multiplex of the program to update.

program_name: The name of the program to update.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"multiplexProgramSettings": The new settings for a multiplex program.

Main.Medialive.update_reservation — Method

update_reservation(reservation_id)
update_reservation(reservation_id, params::Dict{String,<:Any})

Update reservation.

Arguments

reservation_id: Unique reservation ID, e.g. '1234567'

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"name": Name of the reservation

"renewalSettings": Renewal settings for the reservation