Omics

This page documents function available when using the Omics module, created with @service Omics.

Index

Main.Omics.batch_delete_read_set

Main.Omics.cancel_annotation_import_job

Main.Omics.cancel_run

Main.Omics.cancel_variant_import_job

Main.Omics.create_annotation_store

Main.Omics.create_reference_store

Main.Omics.create_run_group

Main.Omics.create_sequence_store

Main.Omics.create_variant_store

Main.Omics.create_workflow

Main.Omics.delete_annotation_store

Main.Omics.delete_reference

Main.Omics.delete_reference_store

Main.Omics.delete_run

Main.Omics.delete_run_group

Main.Omics.delete_sequence_store

Main.Omics.delete_variant_store

Main.Omics.delete_workflow

Main.Omics.get_annotation_import_job

Main.Omics.get_annotation_store

Main.Omics.get_read_set

Main.Omics.get_read_set_activation_job

Main.Omics.get_read_set_export_job

Main.Omics.get_read_set_import_job

Main.Omics.get_read_set_metadata

Main.Omics.get_reference

Main.Omics.get_reference_import_job

Main.Omics.get_reference_metadata

Main.Omics.get_reference_store

Main.Omics.get_run

Main.Omics.get_run_group

Main.Omics.get_run_task

Main.Omics.get_sequence_store

Main.Omics.get_variant_import_job

Main.Omics.get_variant_store

Main.Omics.get_workflow

Main.Omics.list_annotation_import_jobs

Main.Omics.list_annotation_stores

Main.Omics.list_read_set_activation_jobs

Main.Omics.list_read_set_export_jobs

Main.Omics.list_read_set_import_jobs

Main.Omics.list_read_sets

Main.Omics.list_reference_import_jobs

Main.Omics.list_reference_stores

Main.Omics.list_references

Main.Omics.list_run_groups

Main.Omics.list_run_tasks

Main.Omics.list_runs

Main.Omics.list_sequence_stores

Main.Omics.list_tags_for_resource

Main.Omics.list_variant_import_jobs

Main.Omics.list_variant_stores

Main.Omics.list_workflows

Main.Omics.start_annotation_import_job

Main.Omics.start_read_set_activation_job

Main.Omics.start_read_set_export_job

Main.Omics.start_read_set_import_job

Main.Omics.start_reference_import_job

Main.Omics.start_run

Main.Omics.start_variant_import_job

Main.Omics.tag_resource

Main.Omics.untag_resource

Main.Omics.update_annotation_store

Main.Omics.update_run_group

Main.Omics.update_variant_store

Main.Omics.update_workflow

Documentation

Main.Omics.batch_delete_read_set — Method

batch_delete_read_set(ids, sequence_store_id)
batch_delete_read_set(ids, sequence_store_id, params::Dict{String,<:Any})

Deletes one or more read sets.

Arguments

ids: The read sets' IDs.

sequence_store_id: The read sets' sequence store ID.

Main.Omics.cancel_annotation_import_job — Method

cancel_annotation_import_job(job_id)
cancel_annotation_import_job(job_id, params::Dict{String,<:Any})

Cancels an annotation import job.

Arguments

job_id: The job's ID.

Main.Omics.cancel_run — Method

cancel_run(id)
cancel_run(id, params::Dict{String,<:Any})

Cancels a run.

Arguments

id: The run's ID.

Main.Omics.cancel_variant_import_job — Method

cancel_variant_import_job(job_id)
cancel_variant_import_job(job_id, params::Dict{String,<:Any})

Cancels a variant import job.

Arguments

job_id: The job's ID.

Main.Omics.create_annotation_store — Method

create_annotation_store(store_format)
create_annotation_store(store_format, params::Dict{String,<:Any})

Creates an annotation store.

Arguments

store_format: The annotation file format of the store.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A description for the store.

"name": A name for the store.

"reference": The genome reference for the store's annotations.

"sseConfig": Server-side encryption (SSE) settings for the store.

"storeOptions": File parsing options for the annotation store.

"tags": Tags for the store.

Main.Omics.create_reference_store — Method

create_reference_store(name)
create_reference_store(name, params::Dict{String,<:Any})

Creates a reference store.

Arguments

name: A name for the store.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": To ensure that requests don't run multiple times, specify a unique token for each request.

"description": A description for the store.

"sseConfig": Server-side encryption (SSE) settings for the store.

"tags": Tags for the store.

Main.Omics.create_run_group — Method

create_run_group(request_id)
create_run_group(request_id, params::Dict{String,<:Any})

Creates a run group.

Arguments

request_id: A request ID for the group.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxCpus": The maximum number of CPUs to use in the group.

"maxDuration": A max duration for the group.

"maxRuns": The maximum number of concurrent runs for the group.

"name": A name for the group.

"tags": Tags for the group.

Main.Omics.create_sequence_store — Method

create_sequence_store(name)
create_sequence_store(name, params::Dict{String,<:Any})

Creates a sequence store.

Arguments

name: A name for the store.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": To ensure that requests don't run multiple times, specify a unique token for each request.

"description": A description for the store.

"sseConfig": Server-side encryption (SSE) settings for the store.

"tags": Tags for the store.

Main.Omics.create_variant_store — Method

create_variant_store(reference)
create_variant_store(reference, params::Dict{String,<:Any})

Creates a variant store.

Arguments

reference: The genome reference for the store's variants.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A description for the store.

"name": A name for the store.

"sseConfig": Server-side encryption (SSE) settings for the store.

"tags": Tags for the store.

Main.Omics.create_workflow — Method

create_workflow(request_id)
create_workflow(request_id, params::Dict{String,<:Any})

Creates a workflow.

Arguments

request_id: A request ID for the workflow.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"definitionUri": The URI of a definition for the workflow.

"definitionZip": A ZIP archive for the workflow.

"description": A description for the workflow.

"engine": An engine for the workflow.

"main": The path of the main definition file for the workflow.

"name": A name for the workflow.

"parameterTemplate": A parameter template for the workflow.

"storageCapacity": A storage capacity for the workflow.

"tags": Tags for the workflow.

Main.Omics.delete_annotation_store — Method

delete_annotation_store(name)
delete_annotation_store(name, params::Dict{String,<:Any})

Deletes an annotation store.

Arguments

name: The store's name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"force": Whether to force deletion.

Main.Omics.delete_reference — Method

delete_reference(id, reference_store_id)
delete_reference(id, reference_store_id, params::Dict{String,<:Any})

Deletes a genome reference.

Arguments

id: The reference's ID.

reference_store_id: The reference's store ID.

Main.Omics.delete_reference_store — Method

delete_reference_store(id)
delete_reference_store(id, params::Dict{String,<:Any})

Deletes a genome reference store.

Arguments

id: The store's ID.

Main.Omics.delete_run — Method

delete_run(id)
delete_run(id, params::Dict{String,<:Any})

Deletes a workflow run.

Arguments

id: The run's ID.

Main.Omics.delete_run_group — Method

delete_run_group(id)
delete_run_group(id, params::Dict{String,<:Any})

Deletes a workflow run group.

Arguments

id: The run group's ID.

Main.Omics.delete_sequence_store — Method

delete_sequence_store(id)
delete_sequence_store(id, params::Dict{String,<:Any})

Deletes a sequence store.

Arguments

id: The sequence store's ID.

Main.Omics.delete_variant_store — Method

delete_variant_store(name)
delete_variant_store(name, params::Dict{String,<:Any})

Deletes a variant store.

Arguments

name: The store's name.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"force": Whether to force deletion.

Main.Omics.delete_workflow — Method

delete_workflow(id)
delete_workflow(id, params::Dict{String,<:Any})

Deletes a workflow.

Arguments

id: The workflow's ID.

Main.Omics.get_annotation_import_job — Method

get_annotation_import_job(job_id)
get_annotation_import_job(job_id, params::Dict{String,<:Any})

Gets information about an annotation import job.

Arguments

job_id: The job's ID.

Main.Omics.get_annotation_store — Method

get_annotation_store(name)
get_annotation_store(name, params::Dict{String,<:Any})

Gets information about an annotation store.

Arguments

name: The store's name.

Main.Omics.get_read_set — Method

get_read_set(id, part_number, sequence_store_id)
get_read_set(id, part_number, sequence_store_id, params::Dict{String,<:Any})

Gets a file from a read set.

Arguments

id: The read set's ID.

part_number: The part number to retrieve.

sequence_store_id: The read set's sequence store ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"file": The file to retrieve.

Main.Omics.get_read_set_activation_job — Method

get_read_set_activation_job(id, sequence_store_id)
get_read_set_activation_job(id, sequence_store_id, params::Dict{String,<:Any})

Gets information about a read set activation job.

Arguments

id: The job's ID.

sequence_store_id: The job's sequence store ID.

Main.Omics.get_read_set_export_job — Method

get_read_set_export_job(id, sequence_store_id)
get_read_set_export_job(id, sequence_store_id, params::Dict{String,<:Any})

Gets information about a read set export job.

Arguments

id: The job's ID.

sequence_store_id: The job's sequence store ID.

Main.Omics.get_read_set_import_job — Method

get_read_set_import_job(id, sequence_store_id)
get_read_set_import_job(id, sequence_store_id, params::Dict{String,<:Any})

Gets information about a read set import job.

Arguments

id: The job's ID.

sequence_store_id: The job's sequence store ID.

Main.Omics.get_read_set_metadata — Method

get_read_set_metadata(id, sequence_store_id)
get_read_set_metadata(id, sequence_store_id, params::Dict{String,<:Any})

Gets details about a read set.

Arguments

id: The read set's ID.

sequence_store_id: The read set's sequence store ID.

Main.Omics.get_reference — Method

get_reference(id, part_number, reference_store_id)
get_reference(id, part_number, reference_store_id, params::Dict{String,<:Any})

Gets a reference file.

Arguments

id: The reference's ID.

part_number: The part number to retrieve.

reference_store_id: The reference's store ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Range": The range to retrieve.

"file": The file to retrieve.

Main.Omics.get_reference_import_job — Method

get_reference_import_job(id, reference_store_id)
get_reference_import_job(id, reference_store_id, params::Dict{String,<:Any})

Gets information about a reference import job.

Arguments

id: The job's ID.

reference_store_id: The job's reference store ID.

Main.Omics.get_reference_metadata — Method

get_reference_metadata(id, reference_store_id)
get_reference_metadata(id, reference_store_id, params::Dict{String,<:Any})

Gets information about a genome reference's metadata.

Arguments

id: The reference's ID.

reference_store_id: The reference's reference store ID.

Main.Omics.get_reference_store — Method

get_reference_store(id)
get_reference_store(id, params::Dict{String,<:Any})

Gets information about a reference store.

Arguments

id: The store's ID.

Main.Omics.get_run — Method

get_run(id)
get_run(id, params::Dict{String,<:Any})

Gets information about a workflow run.

Arguments

id: The run's ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"export": The run's export format.

Main.Omics.get_run_group — Method

get_run_group(id)
get_run_group(id, params::Dict{String,<:Any})

Gets information about a workflow run group.

Arguments

id: The group's ID.

Main.Omics.get_run_task — Method

get_run_task(id, task_id)
get_run_task(id, task_id, params::Dict{String,<:Any})

Gets information about a workflow run task.

Arguments

id: The task's ID.

task_id: The task's ID.

Main.Omics.get_sequence_store — Method

get_sequence_store(id)
get_sequence_store(id, params::Dict{String,<:Any})

Gets information about a sequence store.

Arguments

id: The store's ID.

Main.Omics.get_variant_import_job — Method

get_variant_import_job(job_id)
get_variant_import_job(job_id, params::Dict{String,<:Any})

Gets information about a variant import job.

Arguments

job_id: The job's ID.

Main.Omics.get_variant_store — Method

get_variant_store(name)
get_variant_store(name, params::Dict{String,<:Any})

Gets information about a variant store.

Arguments

name: The store's name.

Main.Omics.get_workflow — Method

get_workflow(id)
get_workflow(id, params::Dict{String,<:Any})

Gets information about a workflow.

Arguments

id: The workflow's ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"export": The export format for the workflow.

"type": The workflow's type.

Main.Omics.list_annotation_import_jobs — Method

list_annotation_import_jobs()
list_annotation_import_jobs(params::Dict{String,<:Any})

Retrieves a list of annotation import jobs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filter": A filter to apply to the list.

"ids": IDs of annotation import jobs to retrieve.

"maxResults": The maximum number of jobs to return in one page of results.

"nextToken": Specify the pagination token from a previous request to retrieve the next page of results.

Main.Omics.list_annotation_stores — Method

list_annotation_stores()
list_annotation_stores(params::Dict{String,<:Any})

Retrieves a list of annotation stores.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filter": A filter to apply to the list.

"ids": IDs of stores to list.

"maxResults": The maximum number of stores to return in one page of results.

"nextToken": Specify the pagination token from a previous request to retrieve the next page of results.

Main.Omics.list_read_set_activation_jobs — Method

list_read_set_activation_jobs(sequence_store_id)
list_read_set_activation_jobs(sequence_store_id, params::Dict{String,<:Any})

Retrieves a list of read set activation jobs.

Arguments

sequence_store_id: The read set's sequence store ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filter": A filter to apply to the list.

"maxResults": The maximum number of read set activation jobs to return in one page of results.

"nextToken": Specify the pagination token from a previous request to retrieve the next page of results.

Main.Omics.list_read_set_export_jobs — Method

list_read_set_export_jobs(sequence_store_id)
list_read_set_export_jobs(sequence_store_id, params::Dict{String,<:Any})

Retrieves a list of read set export jobs.

Arguments

sequence_store_id: The jobs' sequence store ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filter": A filter to apply to the list.

"maxResults": The maximum number of jobs to return in one page of results.

"nextToken": Specify the pagination token from a previous request to retrieve the next page of results.

Main.Omics.list_read_set_import_jobs — Method

list_read_set_import_jobs(sequence_store_id)
list_read_set_import_jobs(sequence_store_id, params::Dict{String,<:Any})

Retrieves a list of read set import jobs.

Arguments

sequence_store_id: The jobs' sequence store ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filter": A filter to apply to the list.

"maxResults": The maximum number of jobs to return in one page of results.

"nextToken": Specify the pagination token from a previous request to retrieve the next page of results.

Main.Omics.list_read_sets — Method

list_read_sets(sequence_store_id)
list_read_sets(sequence_store_id, params::Dict{String,<:Any})

Retrieves a list of read sets.

Arguments

sequence_store_id: The jobs' sequence store ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filter": A filter to apply to the list.

"maxResults": The maximum number of read sets to return in one page of results.

"nextToken": Specify the pagination token from a previous request to retrieve the next page of results.

Main.Omics.list_reference_import_jobs — Method

list_reference_import_jobs(reference_store_id)
list_reference_import_jobs(reference_store_id, params::Dict{String,<:Any})

Retrieves a list of reference import jobs.

Arguments

reference_store_id: The job's reference store ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filter": A filter to apply to the list.

"maxResults": The maximum number of jobs to return in one page of results.

"nextToken": Specify the pagination token from a previous request to retrieve the next page of results.

Main.Omics.list_reference_stores — Method

list_reference_stores()
list_reference_stores(params::Dict{String,<:Any})

Retrieves a list of reference stores.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filter": A filter to apply to the list.

"maxResults": The maximum number of stores to return in one page of results.

"nextToken": Specify the pagination token from a previous request to retrieve the next page of results.

Main.Omics.list_references — Method

list_references(reference_store_id)
list_references(reference_store_id, params::Dict{String,<:Any})

Retrieves a list of references.

Arguments

reference_store_id: The references' reference store ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filter": A filter to apply to the list.

"maxResults": The maximum number of references to return in one page of results.

"nextToken": Specify the pagination token from a previous request to retrieve the next page of results.

Main.Omics.list_run_groups — Method

list_run_groups()
list_run_groups(params::Dict{String,<:Any})

Retrieves a list of run groups.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of run groups to return in one page of results.

"name": The run groups' name.

"startingToken": Specify the pagination token from a previous request to retrieve the next page of results.

Main.Omics.list_run_tasks — Method

list_run_tasks(id)
list_run_tasks(id, params::Dict{String,<:Any})

Retrieves a list of tasks for a run.

Arguments

id: The run's ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of run tasks to return in one page of results.

"startingToken": Specify the pagination token from a previous request to retrieve the next page of results.

"status": Filter the list by status.

Main.Omics.list_runs — Method

list_runs()
list_runs(params::Dict{String,<:Any})

Retrieves a list of runs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of runs to return in one page of results.

"name": Filter the list by run name.

"runGroupId": Filter the list by run group ID.

"startingToken": Specify the pagination token from a previous request to retrieve the next page of results.

Main.Omics.list_sequence_stores — Method

list_sequence_stores()
list_sequence_stores(params::Dict{String,<:Any})

Retrieves a list of sequence stores.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filter": A filter to apply to the list.

"maxResults": The maximum number of stores to return in one page of results.

"nextToken": Specify the pagination token from a previous request to retrieve the next page of results.

Main.Omics.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Retrieves a list of tags for a resource.

Arguments

resource_arn: The resource's ARN.

Main.Omics.list_variant_import_jobs — Method

list_variant_import_jobs()
list_variant_import_jobs(params::Dict{String,<:Any})

Retrieves a list of variant import jobs.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filter": A filter to apply to the list.

"ids": A list of job IDs.

"maxResults": The maximum number of import jobs to return in one page of results.

"nextToken": Specify the pagination token from a previous request to retrieve the next page of results.

Main.Omics.list_variant_stores — Method

list_variant_stores()
list_variant_stores(params::Dict{String,<:Any})

Retrieves a list of variant stores.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"filter": A filter to apply to the list.

"ids": A list of store IDs.

"maxResults": The maximum number of stores to return in one page of results.

"nextToken": Specify the pagination token from a previous request to retrieve the next page of results.

Main.Omics.list_workflows — Method

list_workflows()
list_workflows(params::Dict{String,<:Any})

Retrieves a list of workflows.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of workflows to return in one page of results.

"name": The workflows' name.

"startingToken": Specify the pagination token from a previous request to retrieve the next page of results.

"type": The workflows' type.

Main.Omics.start_annotation_import_job — Method

start_annotation_import_job(destination_name, items, role_arn)
start_annotation_import_job(destination_name, items, role_arn, params::Dict{String,<:Any})

Starts an annotation import job.

Arguments

destination_name: A destination annotation store for the job.

items: Items to import.

role_arn: A service role for the job.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"formatOptions": Formatting options for the annotation file.

"runLeftNormalization": The job's left normalization setting.

Main.Omics.start_read_set_activation_job — Method

start_read_set_activation_job(sequence_store_id, sources)
start_read_set_activation_job(sequence_store_id, sources, params::Dict{String,<:Any})

Starts a read set activation job.

Arguments

sequence_store_id: The read set's sequence store ID.

sources: The job's sources.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": To ensure that jobs don't run multiple times, specify a unique token for each job.

Main.Omics.start_read_set_export_job — Method

start_read_set_export_job(destination, role_arn, sequence_store_id, sources)
start_read_set_export_job(destination, role_arn, sequence_store_id, sources, params::Dict{String,<:Any})

Starts a read set export job.

Arguments

destination: A location for exported files in Amazon S3.

role_arn: A service role for the job.

sequence_store_id: The read set's sequence store ID.

sources: Sources for the job.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": To ensure that jobs don't run multiple times, specify a unique token for each job.

Main.Omics.start_read_set_import_job — Method

start_read_set_import_job(role_arn, sequence_store_id, sources)
start_read_set_import_job(role_arn, sequence_store_id, sources, params::Dict{String,<:Any})

Starts a read set import job.

Arguments

role_arn: A service role for the job.

sequence_store_id: The read set's sequence store ID.

sources: Source files to import.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": To ensure that jobs don't run multiple times, specify a unique token for each job.

Main.Omics.start_reference_import_job — Method

start_reference_import_job(reference_store_id, role_arn, sources)
start_reference_import_job(reference_store_id, role_arn, sources, params::Dict{String,<:Any})

Starts a reference import job.

Arguments

reference_store_id: The job's reference store ID.

role_arn: A service role for the job.

sources: Sources for the job.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": To ensure that jobs don't run multiple times, specify a unique token for each job.

Main.Omics.start_run — Method

start_run(request_id, role_arn)
start_run(request_id, role_arn, params::Dict{String,<:Any})

Starts a run.

Arguments

request_id: A request ID for the run.

role_arn: A service role for the run.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"logLevel": A log level for the run.

"name": A name for the run.

"outputUri": An output URI for the run.

"parameters": Parameters for the run.

"priority": A priority for the run.

"runGroupId": The run's group ID.

"runId": The run's ID.

"storageCapacity": A storage capacity for the run.

"tags": Tags for the run.

"workflowId": The run's workflow ID.

"workflowType": The run's workflows type.

Main.Omics.start_variant_import_job — Method

start_variant_import_job(destination_name, items, role_arn)
start_variant_import_job(destination_name, items, role_arn, params::Dict{String,<:Any})

Starts a variant import job.

Arguments

destination_name: The destination variant store for the job.

items: Items to import.

role_arn: A service role for the job.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"runLeftNormalization": The job's left normalization setting.

Main.Omics.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Tags a resource.

Arguments

resource_arn: The resource's ARN.

tags: Tags for the resource.

Main.Omics.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes tags from a resource.

Arguments

resource_arn: The resource's ARN.

tag_keys: Keys of tags to remove.

Main.Omics.update_annotation_store — Method

update_annotation_store(name)
update_annotation_store(name, params::Dict{String,<:Any})

Updates an annotation store.

Arguments

name: A name for the store.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A description for the store.

Main.Omics.update_run_group — Method

update_run_group(id)
update_run_group(id, params::Dict{String,<:Any})

Updates a run group.

Arguments

id: The group's ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxCpus": The maximum number of CPUs to use.

"maxDuration": The maximum amount of time to run.

"maxRuns": The maximum number of concurrent runs for the group.

"name": A name for the group.

Main.Omics.update_variant_store — Method

update_variant_store(name)
update_variant_store(name, params::Dict{String,<:Any})

Updates a variant store.

Arguments

name: A name for the store.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A description for the store.

Main.Omics.update_workflow — Method

update_workflow(id)
update_workflow(id, params::Dict{String,<:Any})

Updates a workflow.

Arguments

id: The workflow's ID.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"description": A description for the workflow.

"name": A name for the workflow.