Ram

This page documents function available when using the Ram module, created with @service Ram.

Index

Main.Ram.accept_resource_share_invitation

Main.Ram.associate_resource_share

Main.Ram.associate_resource_share_permission

Main.Ram.create_resource_share

Main.Ram.delete_resource_share

Main.Ram.disassociate_resource_share

Main.Ram.disassociate_resource_share_permission

Main.Ram.enable_sharing_with_aws_organization

Main.Ram.get_permission

Main.Ram.get_resource_policies

Main.Ram.get_resource_share_associations

Main.Ram.get_resource_share_invitations

Main.Ram.get_resource_shares

Main.Ram.list_pending_invitation_resources

Main.Ram.list_permissions

Main.Ram.list_principals

Main.Ram.list_resource_share_permissions

Main.Ram.list_resource_types

Main.Ram.list_resources

Main.Ram.promote_resource_share_created_from_policy

Main.Ram.reject_resource_share_invitation

Main.Ram.tag_resource

Main.Ram.untag_resource

Main.Ram.update_resource_share

Documentation

Main.Ram.accept_resource_share_invitation — Method

accept_resource_share_invitation(resource_share_invitation_arn)
accept_resource_share_invitation(resource_share_invitation_arn, params::Dict{String,<:Any})

Accepts an invitation to a resource share from another AWS account.

Arguments

resource_share_invitation_arn: The Amazon Resource Name (ARN) of the invitation.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

Main.Ram.associate_resource_share — Method

associate_resource_share(resource_share_arn)
associate_resource_share(resource_share_arn, params::Dict{String,<:Any})

Associates the specified resource share with the specified principals and resources.

Arguments

resource_share_arn: The Amazon Resource Name (ARN) of the resource share.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

"principals": The principals to associate with the resource share. The possible values are IDs of AWS accounts, and the ARNs of organizational units (OU) or organizations from AWS Organizations.

"resourceArns": The Amazon Resource Names (ARN) of the resources.

Main.Ram.associate_resource_share_permission — Method

associate_resource_share_permission(permission_arn, resource_share_arn)
associate_resource_share_permission(permission_arn, resource_share_arn, params::Dict{String,<:Any})

Associates a permission with a resource share.

Arguments

permission_arn: The ARN of the AWS RAM permission to associate with the resource share.

resource_share_arn: The Amazon Resource Name (ARN) of the resource share.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

"replace": Indicates whether the permission should replace the permissions that are currently associated with the resource share. Use true to replace the current permissions. Use false to add the permission to the current permission.

Main.Ram.create_resource_share — Method

create_resource_share(name)
create_resource_share(name, params::Dict{String,<:Any})

Creates a resource share.

Arguments

name: The name of the resource share.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"allowExternalPrincipals": Indicates whether principals outside your AWS organization can be associated with a resource share.

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

"permissionArns": The ARNs of the permissions to associate with the resource share. If you do not specify an ARN for the permission, AWS RAM automatically attaches the default version of the permission for each resource type.

"principals": The principals to associate with the resource share. The possible values are IDs of AWS accounts, the ARN of an OU or organization from AWS Organizations.

"resourceArns": The Amazon Resource Names (ARN) of the resources to associate with the resource share.

"tags": One or more tags.

Main.Ram.delete_resource_share — Method

delete_resource_share(resource_share_arn)
delete_resource_share(resource_share_arn, params::Dict{String,<:Any})

Deletes the specified resource share.

Arguments

resource_share_arn: The Amazon Resource Name (ARN) of the resource share.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

Main.Ram.disassociate_resource_share — Method

disassociate_resource_share(resource_share_arn)
disassociate_resource_share(resource_share_arn, params::Dict{String,<:Any})

Disassociates the specified principals or resources from the specified resource share.

Arguments

resource_share_arn: The Amazon Resource Name (ARN) of the resource share.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

"principals": The principals.

"resourceArns": The Amazon Resource Names (ARNs) of the resources.

Main.Ram.disassociate_resource_share_permission — Method

disassociate_resource_share_permission(permission_arn, resource_share_arn)
disassociate_resource_share_permission(permission_arn, resource_share_arn, params::Dict{String,<:Any})

Disassociates an AWS RAM permission from a resource share.

Arguments

permission_arn: The ARN of the permission to disassociate from the resource share.

resource_share_arn: The Amazon Resource Name (ARN) of the resource share.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

Main.Ram.enable_sharing_with_aws_organization — Method

enable_sharing_with_aws_organization()
enable_sharing_with_aws_organization(params::Dict{String,<:Any})

Enables resource sharing within your AWS Organization. The caller must be the master account for the AWS Organization.

Main.Ram.get_permission — Method

get_permission(permission_arn)
get_permission(permission_arn, params::Dict{String,<:Any})

Gets the contents of an AWS RAM permission in JSON format.

Arguments

permission_arn: The ARN of the permission.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"permissionVersion": The identifier for the version of the permission.

Main.Ram.get_resource_policies — Method

get_resource_policies(resource_arns)
get_resource_policies(resource_arns, params::Dict{String,<:Any})

Gets the policies for the specified resources that you own and have shared.

Arguments

resource_arns: The Amazon Resource Names (ARN) of the resources.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

"nextToken": The token for the next page of results.

"principal": The principal.

Main.Ram.get_resource_share_associations — Method

get_resource_share_associations(association_type)
get_resource_share_associations(association_type, params::Dict{String,<:Any})

Gets the resources or principals for the resource shares that you own.

Arguments

association_type: The association type. Specify PRINCIPAL to list the principals that are associated with the specified resource share. Specify RESOURCE to list the resources that are associated with the specified resource share.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"associationStatus": The association status.

"maxResults": The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

"nextToken": The token for the next page of results.

"principal": The principal. You cannot specify this parameter if the association type is RESOURCE.

"resourceArn": The Amazon Resource Name (ARN) of the resource. You cannot specify this parameter if the association type is PRINCIPAL.

"resourceShareArns": The Amazon Resource Names (ARN) of the resource shares.

Main.Ram.get_resource_share_invitations — Method

get_resource_share_invitations()
get_resource_share_invitations(params::Dict{String,<:Any})

Gets the invitations for resource sharing that you've received.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

"nextToken": The token for the next page of results.

"resourceShareArns": The Amazon Resource Names (ARN) of the resource shares.

"resourceShareInvitationArns": The Amazon Resource Names (ARN) of the invitations.

Main.Ram.get_resource_shares — Method

get_resource_shares(resource_owner)
get_resource_shares(resource_owner, params::Dict{String,<:Any})

Gets the resource shares that you own or the resource shares that are shared with you.

Arguments

resource_owner: The type of owner.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

"name": The name of the resource share.

"nextToken": The token for the next page of results.

"resourceShareArns": The Amazon Resource Names (ARN) of the resource shares.

"resourceShareStatus": The status of the resource share.

"tagFilters": One or more tag filters.

Main.Ram.list_pending_invitation_resources — Method

list_pending_invitation_resources(resource_share_invitation_arn)
list_pending_invitation_resources(resource_share_invitation_arn, params::Dict{String,<:Any})

Lists the resources in a resource share that is shared with you but that the invitation is still pending for.

Arguments

resource_share_invitation_arn: The Amazon Resource Name (ARN) of the invitation.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

"nextToken": The token for the next page of results.

Main.Ram.list_permissions — Method

list_permissions()
list_permissions(params::Dict{String,<:Any})

Lists the AWS RAM permissions.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

"nextToken": The token for the next page of results.

"resourceType": Specifies the resource type for which to list permissions. For example, to list only permissions that apply to EC2 subnets, specify ec2:Subnet.

Main.Ram.list_principals — Method

list_principals(resource_owner)
list_principals(resource_owner, params::Dict{String,<:Any})

Lists the principals that you have shared resources with or that have shared resources with you.

Arguments

resource_owner: The type of owner.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

"nextToken": The token for the next page of results.

"principals": The principals.

"resourceArn": The Amazon Resource Name (ARN) of the resource.

"resourceShareArns": The Amazon Resource Names (ARN) of the resource shares.

"resourceType": The resource type. Valid values: acm-pca:CertificateAuthority | appmesh:Mesh | codebuild:Project | codebuild:ReportGroup | ec2:CapacityReservation | ec2:DedicatedHost | ec2:LocalGatewayRouteTable | ec2:PrefixList | ec2:Subnet | ec2:TrafficMirrorTarget | ec2:TransitGateway | imagebuilder:Component | imagebuilder:Image | imagebuilder:ImageRecipe | imagebuilder:ContainerRecipe | glue:Catalog | glue:Database | glue:Table | license-manager:LicenseConfiguration I network-firewall:FirewallPolicy | network-firewall:StatefulRuleGroup | network-firewall:StatelessRuleGroup | outposts:Outpost | resource-groups:Group | rds:Cluster | route53resolver:FirewallRuleGroup |route53resolver:ResolverQueryLogConfig | route53resolver:ResolverRule

Main.Ram.list_resource_share_permissions — Method

list_resource_share_permissions(resource_share_arn)
list_resource_share_permissions(resource_share_arn, params::Dict{String,<:Any})

Lists the AWS RAM permissions that are associated with a resource share.

Arguments

resource_share_arn: The Amazon Resource Name (ARN) of the resource share.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

"nextToken": The token for the next page of results.

Main.Ram.list_resource_types — Method

list_resource_types()
list_resource_types(params::Dict{String,<:Any})

Lists the shareable resource types supported by AWS RAM.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

"nextToken": The token for the next page of results.

Main.Ram.list_resources — Method

list_resources(resource_owner)
list_resources(resource_owner, params::Dict{String,<:Any})

Lists the resources that you added to a resource shares or the resources that are shared with you.

Arguments

resource_owner: The type of owner.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to return with a single call. To retrieve the remaining results, make another call with the returned nextToken value.

"nextToken": The token for the next page of results.

"principal": The principal.

"resourceArns": The Amazon Resource Names (ARN) of the resources.

"resourceShareArns": The Amazon Resource Names (ARN) of the resource shares.

"resourceType": The resource type. Valid values: acm-pca:CertificateAuthority | appmesh:Mesh | codebuild:Project | codebuild:ReportGroup | ec2:CapacityReservation | ec2:DedicatedHost | ec2:LocalGatewayRouteTable | ec2:PrefixList | ec2:Subnet | ec2:TrafficMirrorTarget | ec2:TransitGateway | imagebuilder:Component | imagebuilder:Image | imagebuilder:ImageRecipe | imagebuilder:ContainerRecipe | glue:Catalog | glue:Database | glue:Table | license-manager:LicenseConfiguration I network-firewall:FirewallPolicy | network-firewall:StatefulRuleGroup | network-firewall:StatelessRuleGroup | outposts:Outpost | resource-groups:Group | rds:Cluster | route53resolver:FirewallRuleGroup |route53resolver:ResolverQueryLogConfig | route53resolver:ResolverRule

Main.Ram.promote_resource_share_created_from_policy — Method

promote_resource_share_created_from_policy(resource_share_arn)
promote_resource_share_created_from_policy(resource_share_arn, params::Dict{String,<:Any})

Resource shares that were created by attaching a policy to a resource are visible only to the resource share owner, and the resource share cannot be modified in AWS RAM. Use this API action to promote the resource share. When you promote the resource share, it becomes: Visible to all principals that it is shared with. Modifiable in AWS RAM.

Arguments

resource_share_arn: The ARN of the resource share to promote.

Main.Ram.reject_resource_share_invitation — Method

reject_resource_share_invitation(resource_share_invitation_arn)
reject_resource_share_invitation(resource_share_invitation_arn, params::Dict{String,<:Any})

Rejects an invitation to a resource share from another AWS account.

Arguments

resource_share_invitation_arn: The Amazon Resource Name (ARN) of the invitation.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

Main.Ram.tag_resource — Method

tag_resource(resource_share_arn, tags)
tag_resource(resource_share_arn, tags, params::Dict{String,<:Any})

Adds the specified tags to the specified resource share that you own.

Arguments

resource_share_arn: The Amazon Resource Name (ARN) of the resource share.

tags: One or more tags.

Main.Ram.untag_resource — Method

untag_resource(resource_share_arn, tag_keys)
untag_resource(resource_share_arn, tag_keys, params::Dict{String,<:Any})

Removes the specified tags from the specified resource share that you own.

Arguments

resource_share_arn: The Amazon Resource Name (ARN) of the resource share.

tag_keys: The tag keys of the tags to remove.

Main.Ram.update_resource_share — Method

update_resource_share(resource_share_arn)
update_resource_share(resource_share_arn, params::Dict{String,<:Any})

Updates the specified resource share that you own.

Arguments

resource_share_arn: The Amazon Resource Name (ARN) of the resource share.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"allowExternalPrincipals": Indicates whether principals outside your AWS organization can be associated with a resource share.

"clientToken": A unique, case-sensitive identifier that you provide to ensure the idempotency of the request.

"name": The name of the resource share.