Resiliencehub

This page documents function available when using the Resiliencehub module, created with @service Resiliencehub.

Index

Main.Resiliencehub.add_draft_app_version_resource_mappings

Main.Resiliencehub.create_app

Main.Resiliencehub.create_app_version_app_component

Main.Resiliencehub.create_app_version_resource

Main.Resiliencehub.create_recommendation_template

Main.Resiliencehub.create_resiliency_policy

Main.Resiliencehub.delete_app

Main.Resiliencehub.delete_app_assessment

Main.Resiliencehub.delete_app_input_source

Main.Resiliencehub.delete_app_version_app_component

Main.Resiliencehub.delete_app_version_resource

Main.Resiliencehub.delete_recommendation_template

Main.Resiliencehub.delete_resiliency_policy

Main.Resiliencehub.describe_app

Main.Resiliencehub.describe_app_assessment

Main.Resiliencehub.describe_app_version

Main.Resiliencehub.describe_app_version_app_component

Main.Resiliencehub.describe_app_version_resource

Main.Resiliencehub.describe_app_version_resources_resolution_status

Main.Resiliencehub.describe_app_version_template

Main.Resiliencehub.describe_draft_app_version_resources_import_status

Main.Resiliencehub.describe_resiliency_policy

Main.Resiliencehub.import_resources_to_draft_app_version

Main.Resiliencehub.list_alarm_recommendations

Main.Resiliencehub.list_app_assessments

Main.Resiliencehub.list_app_component_compliances

Main.Resiliencehub.list_app_component_recommendations

Main.Resiliencehub.list_app_input_sources

Main.Resiliencehub.list_app_version_app_components

Main.Resiliencehub.list_app_version_resource_mappings

Main.Resiliencehub.list_app_version_resources

Main.Resiliencehub.list_app_versions

Main.Resiliencehub.list_apps

Main.Resiliencehub.list_recommendation_templates

Main.Resiliencehub.list_resiliency_policies

Main.Resiliencehub.list_sop_recommendations

Main.Resiliencehub.list_suggested_resiliency_policies

Main.Resiliencehub.list_tags_for_resource

Main.Resiliencehub.list_test_recommendations

Main.Resiliencehub.list_unsupported_app_version_resources

Main.Resiliencehub.publish_app_version

Main.Resiliencehub.put_draft_app_version_template

Main.Resiliencehub.remove_draft_app_version_resource_mappings

Main.Resiliencehub.resolve_app_version_resources

Main.Resiliencehub.start_app_assessment

Main.Resiliencehub.tag_resource

Main.Resiliencehub.untag_resource

Main.Resiliencehub.update_app

Main.Resiliencehub.update_app_version

Main.Resiliencehub.update_app_version_app_component

Main.Resiliencehub.update_app_version_resource

Main.Resiliencehub.update_resiliency_policy

Documentation

Main.Resiliencehub.add_draft_app_version_resource_mappings — Method

add_draft_app_version_resource_mappings(app_arn, resource_mappings)
add_draft_app_version_resource_mappings(app_arn, resource_mappings, params::Dict{String,<:Any})

Adds the resource mapping for the draft application version. You can also update an existing resource mapping to a new physical resource.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

resource_mappings: Mappings used to map logical resources from the template to physical resources. You can use the mapping type CFNSTACK if the application template uses a logical stack name. Or you can map individual resources by using the mapping type RESOURCE. We recommend using the mapping type CFNSTACK if the application is backed by a CloudFormation stack.

Main.Resiliencehub.create_app — Method

create_app(name)
create_app(name, params::Dict{String,<:Any})

Creates an AWS Resilience Hub application. An AWS Resilience Hub application is a collection of Amazon Web Services resources structured to prevent and recover Amazon Web Services application disruptions. To describe a AWS Resilience Hub application, you provide an application name, resources from one or more–up to five–CloudFormation stacks, and an appropriate resiliency policy. After you create an AWS Resilience Hub application, you publish it so that you can run a resiliency assessment on it. You can then use recommendations from the assessment to improve resiliency by running another assessment, comparing results, and then iterating the process until you achieve your goals for recovery time objective (RTO) and recovery point objective (RPO).

Arguments

name: The name for the application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"assessmentSchedule": Assessment execution schedule with 'Daily' or 'Disabled' values.

"clientToken": Used for an idempotency token. A client token is a unique, case-sensitive string of up to 64 ASCII characters. You should not reuse the same client token for other API requests.

"description": The optional description for an app.

"policyArn": The Amazon Resource Name (ARN) of the resiliency policy. The format for this ARN is: arn:partition:resiliencehub:region:account:resiliency-policy/policy-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

"tags": The tags assigned to the resource. A tag is a label that you assign to an Amazon Web Services resource. Each tag consists of a key/value pair.

Main.Resiliencehub.create_app_version_app_component — Method

create_app_version_app_component(app_arn, name, type)
create_app_version_app_component(app_arn, name, type, params::Dict{String,<:Any})

Creates a new Application Component in the AWS Resilience Hub application. This API updates the AWS Resilience Hub application draft version. To use this Application Component for running assessments, you must publish the AWS Resilience Hub application using the PublishAppVersion API.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

name: The name of the Application Component.

type: The type of Application Component. For more information about the types of Application Component, see Grouping resources in an AppComponent.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"additionalInfo": Currently, there is no supported additional information for Application Components.

"clientToken": Used for an idempotency token. A client token is a unique, case-sensitive string of up to 64 ASCII characters. You should not reuse the same client token for other API requests.

"id": The identifier of the Application Component.

Main.Resiliencehub.create_app_version_resource — Method

create_app_version_resource(app_arn, app_components, logical_resource_id, physical_resource_id, resource_name, resource_type)
create_app_version_resource(app_arn, app_components, logical_resource_id, physical_resource_id, resource_name, resource_type, params::Dict{String,<:Any})

Adds a resource to the AWS Resilience Hub applicationand assigns it to the specified Application Components. If you specify a new Application Component, AWS Resilience Hub will automatically create the Application Component. This action has no effect outside AWS Resilience Hub. This API updates the AWS Resilience Hub application draft version. To use this resource for running resiliency assessments, you must publish the AWS Resilience Hub application using the PublishAppVersion API. To update application version with new physicalResourceID, you must call ResolveAppVersionResources API.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

app_components: The list of Application Components that this resource belongs to. If an Application Component is not part of the AWS Resilience Hub application, it will be added.

logical_resource_id: The logical identifier of the resource.

physical_resource_id: The physical identifier of the resource.

resource_name: The name of the resource.

resource_type: The type of resource.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"additionalInfo": Currently, there is no supported additional information for resources.

"awsAccountId": The Amazon Web Services account that owns the physical resource.

"awsRegion": The Amazon Web Services region that owns the physical resource.

"clientToken": Used for an idempotency token. A client token is a unique, case-sensitive string of up to 64 ASCII characters. You should not reuse the same client token for other API requests.

Main.Resiliencehub.create_recommendation_template — Method

create_recommendation_template(assessment_arn, name)
create_recommendation_template(assessment_arn, name, params::Dict{String,<:Any})

Creates a new recommendation template for the AWS Resilience Hub application.

Arguments

assessment_arn: The Amazon Resource Name (ARN) of the assessment. The format for this ARN is: arn:partition:resiliencehub:region:account:app-assessment/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

name: The name for the recommendation template.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"bucketName": The name of the Amazon S3 bucket that will contain the recommendation template.

"clientToken": Used for an idempotency token. A client token is a unique, case-sensitive string of up to 64 ASCII characters. You should not reuse the same client token for other API requests.

"format": The format for the recommendation template. CfnJson The template is CloudFormation JSON. CfnYaml The template is CloudFormation YAML.

"recommendationIds": Identifiers for the recommendations used to create a recommendation template.

"recommendationTypes": An array of strings that specify the recommendation template type or types. Alarm The template is an AlarmRecommendation template. Sop The template is a SopRecommendation template. Test The template is a TestRecommendation template.

"tags": The tags assigned to the resource. A tag is a label that you assign to an Amazon Web Services resource. Each tag consists of a key/value pair.

Main.Resiliencehub.create_resiliency_policy — Method

create_resiliency_policy(policy, policy_name, tier)
create_resiliency_policy(policy, policy_name, tier, params::Dict{String,<:Any})

Creates a resiliency policy for an application.

Arguments

policy: The type of resiliency policy to be created, including the recovery time objective (RTO) and recovery point objective (RPO) in seconds.

policy_name: The name of the policy

tier: The tier for this resiliency policy, ranging from the highest severity (MissionCritical) to lowest (NonCritical).

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Used for an idempotency token. A client token is a unique, case-sensitive string of up to 64 ASCII characters. You should not reuse the same client token for other API requests.

"dataLocationConstraint": Specifies a high-level geographical location constraint for where your resilience policy data can be stored.

"policyDescription": The description for the policy.

"tags": The tags assigned to the resource. A tag is a label that you assign to an Amazon Web Services resource. Each tag consists of a key/value pair.

Main.Resiliencehub.delete_app — Method

delete_app(app_arn)
delete_app(app_arn, params::Dict{String,<:Any})

Deletes an AWS Resilience Hub application. This is a destructive action that can't be undone.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Used for an idempotency token. A client token is a unique, case-sensitive string of up to 64 ASCII characters. You should not reuse the same client token for other API requests.

"forceDelete": A boolean option to force the deletion of an AWS Resilience Hub application.

Main.Resiliencehub.delete_app_assessment — Method

delete_app_assessment(assessment_arn)
delete_app_assessment(assessment_arn, params::Dict{String,<:Any})

Deletes an AWS Resilience Hub application assessment. This is a destructive action that can't be undone.

Arguments

assessment_arn: The Amazon Resource Name (ARN) of the assessment. The format for this ARN is: arn:partition:resiliencehub:region:account:app-assessment/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Used for an idempotency token. A client token is a unique, case-sensitive string of up to 64 ASCII characters. You should not reuse the same client token for other API requests.

Main.Resiliencehub.delete_app_input_source — Method

delete_app_input_source(app_arn)
delete_app_input_source(app_arn, params::Dict{String,<:Any})

Deletes the input source and all of its imported resources from the AWS Resilience Hub application.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Used for an idempotency token. A client token is a unique, case-sensitive string of up to 64 ASCII characters. You should not reuse the same client token for other API requests.

"sourceArn": The Amazon Resource Name (ARN) of the imported resource you want to remove from the AWS Resilience Hub application. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

"terraformSource": The imported Terraform s3 state ﬁle you want to remove from the AWS Resilience Hub application.

Main.Resiliencehub.delete_app_version_app_component — Method

delete_app_version_app_component(app_arn, id)
delete_app_version_app_component(app_arn, id, params::Dict{String,<:Any})

Deletes an Application Component from the AWS Resilience Hub application. This API updates the AWS Resilience Hub application draft version. To use this Application Component for running assessments, you must publish the AWS Resilience Hub application using the PublishAppVersion API. You will not be able to delete an Application Component if it has resources associated with it.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

id: The identifier of the Application Component.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Used for an idempotency token. A client token is a unique, case-sensitive string of up to 64 ASCII characters. You should not reuse the same client token for other API requests.

Main.Resiliencehub.delete_app_version_resource — Method

delete_app_version_resource(app_arn)
delete_app_version_resource(app_arn, params::Dict{String,<:Any})

Deletes a resource from the AWS Resilience Hub application. You can only delete a manually added resource. To exclude non-manually added resources, use the UpdateAppVersionResource API. This action has no effect outside AWS Resilience Hub. This API updates the AWS Resilience Hub application draft version. To use this resource for running resiliency assessments, you must publish the AWS Resilience Hub application using the PublishAppVersion API.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"awsAccountId": The Amazon Web Services account that owns the physical resource.

"awsRegion": The Amazon Web Services region that owns the physical resource.

"clientToken": Used for an idempotency token. A client token is a unique, case-sensitive string of up to 64 ASCII characters. You should not reuse the same client token for other API requests.

"logicalResourceId": The logical identifier of the resource.

"physicalResourceId": The physical identifier of the resource.

"resourceName": The name of the resource.

Main.Resiliencehub.delete_recommendation_template — Method

delete_recommendation_template(recommendation_template_arn)
delete_recommendation_template(recommendation_template_arn, params::Dict{String,<:Any})

Deletes a recommendation template. This is a destructive action that can't be undone.

Arguments

recommendation_template_arn: The Amazon Resource Name (ARN) for a recommendation template.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Used for an idempotency token. A client token is a unique, case-sensitive string of up to 64 ASCII characters. You should not reuse the same client token for other API requests.

Main.Resiliencehub.delete_resiliency_policy — Method

delete_resiliency_policy(policy_arn)
delete_resiliency_policy(policy_arn, params::Dict{String,<:Any})

Deletes a resiliency policy. This is a destructive action that can't be undone.

Arguments

policy_arn: The Amazon Resource Name (ARN) of the resiliency policy. The format for this ARN is: arn:partition:resiliencehub:region:account:resiliency-policy/policy-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Used for an idempotency token. A client token is a unique, case-sensitive string of up to 64 ASCII characters. You should not reuse the same client token for other API requests.

Main.Resiliencehub.describe_app — Method

describe_app(app_arn)
describe_app(app_arn, params::Dict{String,<:Any})

Describes an AWS Resilience Hub application.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

Main.Resiliencehub.describe_app_assessment — Method

describe_app_assessment(assessment_arn)
describe_app_assessment(assessment_arn, params::Dict{String,<:Any})

Describes an assessment for an AWS Resilience Hub application.

Arguments

assessment_arn: The Amazon Resource Name (ARN) of the assessment. The format for this ARN is: arn:partition:resiliencehub:region:account:app-assessment/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

Main.Resiliencehub.describe_app_version — Method

describe_app_version(app_arn, app_version)
describe_app_version(app_arn, app_version, params::Dict{String,<:Any})

Describes the AWS Resilience Hub application version.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

app_version: The AWS Resilience Hub application version.

Main.Resiliencehub.describe_app_version_app_component — Method

describe_app_version_app_component(app_arn, app_version, id)
describe_app_version_app_component(app_arn, app_version, id, params::Dict{String,<:Any})

Describes an Application Component in the AWS Resilience Hub application.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

app_version: The AWS Resilience Hub application version.

id: The identifier of the Application Component.

Main.Resiliencehub.describe_app_version_resource — Method

describe_app_version_resource(app_arn, app_version)
describe_app_version_resource(app_arn, app_version, params::Dict{String,<:Any})

Describes a resource of the AWS Resilience Hub application. This API accepts only one of the following parameters to descibe the resource: resourceName logicalResourceId physicalResourceId (Along with physicalResourceId, you can also provide awsAccountId, and awsRegion)

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

app_version: The AWS Resilience Hub application version.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"awsAccountId": The Amazon Web Services account that owns the physical resource.

"awsRegion": The Amazon Web Services region that owns the physical resource.

"logicalResourceId": The logical identifier of the resource.

"physicalResourceId": The physical identifier of the resource.

"resourceName": The name of the resource.

Main.Resiliencehub.describe_app_version_resources_resolution_status — Method

describe_app_version_resources_resolution_status(app_arn, app_version)
describe_app_version_resources_resolution_status(app_arn, app_version, params::Dict{String,<:Any})

Returns the resolution status for the specified resolution identifier for an application version. If resolutionId is not specified, the current resolution status is returned.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

app_version: The version of the application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"resolutionId": The identifier for a specific resolution.

Main.Resiliencehub.describe_app_version_template — Method

describe_app_version_template(app_arn, app_version)
describe_app_version_template(app_arn, app_version, params::Dict{String,<:Any})

Describes details about an AWS Resilience Hub application.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

app_version: The version of the application.

Main.Resiliencehub.describe_draft_app_version_resources_import_status — Method

describe_draft_app_version_resources_import_status(app_arn)
describe_draft_app_version_resources_import_status(app_arn, params::Dict{String,<:Any})

Describes the status of importing resources to an application version. If you get a 404 error with ResourceImportStatusNotFoundAppMetadataException, you must call importResourcesToDraftAppVersion after creating the application and before calling describeDraftAppVersionResourcesImportStatus to obtain the status.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

Main.Resiliencehub.describe_resiliency_policy — Method

describe_resiliency_policy(policy_arn)
describe_resiliency_policy(policy_arn, params::Dict{String,<:Any})

Describes a specified resiliency policy for an AWS Resilience Hub application. The returned policy object includes creation time, data location constraints, the Amazon Resource Name (ARN) for the policy, tags, tier, and more.

Arguments

policy_arn: The Amazon Resource Name (ARN) of the resiliency policy. The format for this ARN is: arn:partition:resiliencehub:region:account:resiliency-policy/policy-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

Main.Resiliencehub.import_resources_to_draft_app_version — Method

import_resources_to_draft_app_version(app_arn)
import_resources_to_draft_app_version(app_arn, params::Dict{String,<:Any})

Imports resources to AWS Resilience Hub application draft version from different input sources. For more information about the input sources supported by AWS Resilience Hub, see Discover the structure and describe your Resilience Hub application.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"importStrategy": The import strategy you would like to set to import resources into AWS Resilience Hub application.

"sourceArns": The Amazon Resource Names (ARNs) for the resources that you want to import.

"terraformSources": A list of terraform file s3 URLs you need to import.

Main.Resiliencehub.list_alarm_recommendations — Method

list_alarm_recommendations(assessment_arn)
list_alarm_recommendations(assessment_arn, params::Dict{String,<:Any})

Lists the alarm recommendations for an AWS Resilience Hub application.

Arguments

assessment_arn: The Amazon Resource Name (ARN) of the assessment. The format for this ARN is: arn:partition:resiliencehub:region:account:app-assessment/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"nextToken": Null, or the token from a previous call to get the next set of results.

Main.Resiliencehub.list_app_assessments — Method

list_app_assessments()
list_app_assessments(params::Dict{String,<:Any})

Lists the assessments for an AWS Resilience Hub application. You can use request parameters to refine the results for the response object.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"appArn": The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

"assessmentName": The name for the assessment.

"assessmentStatus": The current status of the assessment for the resiliency policy.

"complianceStatus": The current status of compliance for the resiliency policy.

"invoker": Specifies the entity that invoked a specific assessment, either a User or the System.

"maxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"nextToken": Null, or the token from a previous call to get the next set of results.

"reverseOrder": The default is to sort by ascending startTime. To sort by descending startTime, set reverseOrder to true.

Main.Resiliencehub.list_app_component_compliances — Method

list_app_component_compliances(assessment_arn)
list_app_component_compliances(assessment_arn, params::Dict{String,<:Any})

Lists the compliances for an AWS Resilience Hub Application Component.

Arguments

assessment_arn: The Amazon Resource Name (ARN) of the assessment. The format for this ARN is: arn:partition:resiliencehub:region:account:app-assessment/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"nextToken": Null, or the token from a previous call to get the next set of results.

Main.Resiliencehub.list_app_component_recommendations — Method

list_app_component_recommendations(assessment_arn)
list_app_component_recommendations(assessment_arn, params::Dict{String,<:Any})

Lists the recommendations for an AWS Resilience Hub Application Component.

Arguments

assessment_arn: The Amazon Resource Name (ARN) of the assessment. The format for this ARN is: arn:partition:resiliencehub:region:account:app-assessment/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"nextToken": Null, or the token from a previous call to get the next set of results.

Main.Resiliencehub.list_app_input_sources — Method

list_app_input_sources(app_arn, app_version)
list_app_input_sources(app_arn, app_version, params::Dict{String,<:Any})

Lists all the input sources of the AWS Resilience Hub application. For more information about the input sources supported by AWS Resilience Hub, see Discover the structure and describe your Resilience Hub application.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

app_version: The AWS Resilience Hub application version.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Maximum number of input sources to be displayed per AWS Resilience Hub application.

"nextToken": Null, or the token from a previous call to get the next set of results.

Main.Resiliencehub.list_app_version_app_components — Method

list_app_version_app_components(app_arn, app_version)
list_app_version_app_components(app_arn, app_version, params::Dict{String,<:Any})

Lists all the Application Components in the AWS Resilience Hub application.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

app_version: The version of the Application Component.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": Maximum number of Application Components to be displayed per AWS Resilience Hub application version.

"nextToken": Null, or the token from a previous call to get the next set of results.

Main.Resiliencehub.list_app_version_resource_mappings — Method

list_app_version_resource_mappings(app_arn, app_version)
list_app_version_resource_mappings(app_arn, app_version, params::Dict{String,<:Any})

Lists how the resources in an application version are mapped/sourced from. Mappings can be physical resource identifiers, CloudFormation stacks, resource-groups, or an application registry app.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

app_version: The version of the application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"nextToken": Null, or the token from a previous call to get the next set of results.

Main.Resiliencehub.list_app_version_resources — Method

list_app_version_resources(app_arn, app_version)
list_app_version_resources(app_arn, app_version, params::Dict{String,<:Any})

Lists all the resources in an AWS Resilience Hub application.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

app_version: The version of the application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"nextToken": Null, or the token from a previous call to get the next set of results.

"resolutionId": The identifier for a specific resolution.

Main.Resiliencehub.list_app_versions — Method

list_app_versions(app_arn)
list_app_versions(app_arn, params::Dict{String,<:Any})

Lists the different versions for the AWS Resilience Hub applications.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"nextToken": Null, or the token from a previous call to get the next set of results.

Main.Resiliencehub.list_apps — Method

list_apps()
list_apps(params::Dict{String,<:Any})

Lists your AWS Resilience Hub applications. You can filter applications using only one filter at a time or without using any filter. If you try to filter applications using multiple filters, you will get the following error: An error occurred (ValidationException) when calling the ListApps operation: Only one filter is supported for this operation.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"appArn": The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

"maxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"name": The name for the one of the listed applications.

"nextToken": Null, or the token from a previous call to get the next set of results.

Main.Resiliencehub.list_recommendation_templates — Method

list_recommendation_templates(assessment_arn)
list_recommendation_templates(assessment_arn, params::Dict{String,<:Any})

Lists the recommendation templates for the AWS Resilience Hub applications.

Arguments

assessment_arn: The Amazon Resource Name (ARN) of the assessment. The format for this ARN is: arn:partition:resiliencehub:region:account:app-assessment/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"name": The name for one of the listed recommendation templates.

"nextToken": Null, or the token from a previous call to get the next set of results.

"recommendationTemplateArn": The Amazon Resource Name (ARN) for a recommendation template.

"reverseOrder": The default is to sort by ascending startTime. To sort by descending startTime, set reverseOrder to true.

"status": The status of the action.

Main.Resiliencehub.list_resiliency_policies — Method

list_resiliency_policies()
list_resiliency_policies(params::Dict{String,<:Any})

Lists the resiliency policies for the AWS Resilience Hub applications.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"nextToken": Null, or the token from a previous call to get the next set of results.

"policyName": The name of the policy

Main.Resiliencehub.list_sop_recommendations — Method

list_sop_recommendations(assessment_arn)
list_sop_recommendations(assessment_arn, params::Dict{String,<:Any})

Lists the standard operating procedure (SOP) recommendations for the AWS Resilience Hub applications.

Arguments

assessment_arn: The Amazon Resource Name (ARN) of the assessment. The format for this ARN is: arn:partition:resiliencehub:region:account:app-assessment/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"nextToken": Null, or the token from a previous call to get the next set of results.

Main.Resiliencehub.list_suggested_resiliency_policies — Method

list_suggested_resiliency_policies()
list_suggested_resiliency_policies(params::Dict{String,<:Any})

Lists the suggested resiliency policies for the AWS Resilience Hub applications.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"nextToken": Null, or the token from a previous call to get the next set of results.

Main.Resiliencehub.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Lists the tags for your resources in your AWS Resilience Hub applications.

Arguments

resource_arn: The Amazon Resource Name (ARN) for a specific resource in your AWS Resilience Hub application.

Main.Resiliencehub.list_test_recommendations — Method

list_test_recommendations(assessment_arn)
list_test_recommendations(assessment_arn, params::Dict{String,<:Any})

Lists the test recommendations for the AWS Resilience Hub application.

Arguments

assessment_arn: The Amazon Resource Name (ARN) of the assessment. The format for this ARN is: arn:partition:resiliencehub:region:account:app-assessment/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"nextToken": Null, or the token from a previous call to get the next set of results.

Main.Resiliencehub.list_unsupported_app_version_resources — Method

list_unsupported_app_version_resources(app_arn, app_version)
list_unsupported_app_version_resources(app_arn, app_version, params::Dict{String,<:Any})

Lists the resources that are not currently supported in AWS Resilience Hub. An unsupported resource is a resource that exists in the object that was used to create an app, but is not supported by AWS Resilience Hub.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

app_version: The version of the application.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results to include in the response. If more results exist than the specified MaxResults value, a token is included in the response so that the remaining results can be retrieved.

"nextToken": Null, or the token from a previous call to get the next set of results.

"resolutionId": The identifier for a specific resolution.

Main.Resiliencehub.publish_app_version — Method

publish_app_version(app_arn)
publish_app_version(app_arn, params::Dict{String,<:Any})

Publishes a new version of a specific AWS Resilience Hub application.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

Main.Resiliencehub.put_draft_app_version_template — Method

put_draft_app_version_template(app_arn, app_template_body)
put_draft_app_version_template(app_arn, app_template_body, params::Dict{String,<:Any})

Adds or updates the app template for an AWS Resilience Hub application draft version.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

app_template_body: A JSON string that provides information about your application structure. To learn more about the appTemplateBody template, see the sample template provided in the Examples section. The appTemplateBody JSON string has the following structure: resources The list of logical resources that needs to be included in the application. Type: Array Don't add the resources that you want to exclude. Each resources array item includes the following fields: logicalResourceId The logical identifier of the resource. Type: Object Each logicalResourceId object includes the following fields: identifier The identifier of the resource. Type: String logicalStackName The name of the CloudFormation stack this resource belongs to. Type: String resourceGroupName The name of the resource group this resource belongs to. Type: String terraformSourceName The name of the Terraform S3 state file this resource belongs to. Type: String type The type of resource. Type: string name The name of the resource. Type: String appComponents The list of Application Components that this resource belongs to. If an Application Component is not part of the AWS Resilience Hub application, it will be added. Type: Array Each appComponents array item includes the following fields: name The name of the Application Component. Type: String type The type of Application Component. For more information about the types of Application Component, see Grouping resources in an AppComponent. Type: String resourceNames The list of included resources that are assigned to the Application Component. Type: Array of strings additionalInfo Additional configuration parameters for an AWS Resilience Hub application. Currently, this parameter accepts a key-value mapping (in a string format) of only one failover region and one associated account. Key: "failover-regions" Value: "[{"region":"&lt;REGION&gt;", "accounts":[{"id":"&lt;ACCOUNT_ID&gt;"}]}]" excludedResources The list of logical resource identifiers to be excluded from the application. Type: Array Don't add the resources that you want to include. Each excludedResources array item includes the following fields: logicalResourceIds The logical identifier of the resource. Type: Object You can configure only one of the following fields: logicalStackName resourceGroupName terraformSourceName Each logicalResourceIds object includes the following fields: identifier The identifier of the resource. Type: String logicalStackName The name of the CloudFormation stack this resource belongs to. Type: String resourceGroupName The name of the resource group this resource belongs to. Type: String terraformSourceName The name of the Terraform S3 state file this resource belongs to. Type: String

Main.Resiliencehub.remove_draft_app_version_resource_mappings — Method

remove_draft_app_version_resource_mappings(app_arn)
remove_draft_app_version_resource_mappings(app_arn, params::Dict{String,<:Any})

Removes resource mappings from a draft application version.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"appRegistryAppNames": The names of the registered applications you want to remove from the resource mappings.

"logicalStackNames": The names of the CloudFormation stacks you want to remove from the resource mappings.

"resourceGroupNames": The names of the resource groups you want to remove from the resource mappings.

"resourceNames": The names of the resources you want to remove from the resource mappings.

"terraformSourceNames": The names of the Terraform sources you want to remove from the resource mappings.

Main.Resiliencehub.resolve_app_version_resources — Method

resolve_app_version_resources(app_arn, app_version)
resolve_app_version_resources(app_arn, app_version, params::Dict{String,<:Any})

Resolves the resources for an application version.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

app_version: The version of the application.

Main.Resiliencehub.start_app_assessment — Method

start_app_assessment(app_arn, app_version, assessment_name)
start_app_assessment(app_arn, app_version, assessment_name, params::Dict{String,<:Any})

Creates a new application assessment for an application.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

app_version: The version of the application.

assessment_name: The name for the assessment.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"clientToken": Used for an idempotency token. A client token is a unique, case-sensitive string of up to 64 ASCII characters. You should not reuse the same client token for other API requests.

"tags": The tags assigned to the resource. A tag is a label that you assign to an Amazon Web Services resource. Each tag consists of a key/value pair.

Main.Resiliencehub.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Applies one or more tags to a resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

tags: The tags to assign to the resource. Each tag consists of a key/value pair.

Main.Resiliencehub.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes one or more tags from a resource.

Arguments

resource_arn: The Amazon Resource Name (ARN) of the resource.

tag_keys: The keys of the tags you want to remove.

Main.Resiliencehub.update_app — Method

update_app(app_arn)
update_app(app_arn, params::Dict{String,<:Any})

Updates an application.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"assessmentSchedule": Assessment execution schedule with 'Daily' or 'Disabled' values.

"clearResiliencyPolicyArn": Specifies if the resiliency policy ARN should be cleared.

"description": The optional description for an app.

"policyArn": The Amazon Resource Name (ARN) of the resiliency policy. The format for this ARN is: arn:partition:resiliencehub:region:account:resiliency-policy/policy-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

Main.Resiliencehub.update_app_version — Method

update_app_version(app_arn)
update_app_version(app_arn, params::Dict{String,<:Any})

Updates the AWS Resilience Hub application version. This API updates the AWS Resilience Hub application draft version. To use this information for running resiliency assessments, you must publish the AWS Resilience Hub application using the PublishAppVersion API.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"additionalInfo": Additional configuration parameters for an AWS Resilience Hub application. Currently, this parameter accepts a key-value mapping (in a string format) of only one failover region and one associated account. Key: "failover-regions" Value: "[{"region":"&lt;REGION&gt;", "accounts":[{"id":"&lt;ACCOUNT_ID&gt;"}]}]"

Main.Resiliencehub.update_app_version_app_component — Method

update_app_version_app_component(app_arn, id)
update_app_version_app_component(app_arn, id, params::Dict{String,<:Any})

Updates an existing Application Component in the AWS Resilience Hub application. This API updates the AWS Resilience Hub application draft version. To use this Application Component for running assessments, you must publish the AWS Resilience Hub application using the PublishAppVersion API.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

id: The identifier of the Application Component.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"additionalInfo": Currently, there is no supported additional information for Application Components.

"name": The name of the Application Component.

"type": The type of Application Component. For more information about the types of Application Component, see Grouping resources in an AppComponent.

Main.Resiliencehub.update_app_version_resource — Method

update_app_version_resource(app_arn)
update_app_version_resource(app_arn, params::Dict{String,<:Any})

Updates the resource details in the AWS Resilience Hub application. This action has no effect outside AWS Resilience Hub. This API updates the AWS Resilience Hub application draft version. To use this resource for running resiliency assessments, you must publish the AWS Resilience Hub application using the PublishAppVersion API. To update application version with new physicalResourceID, you must call ResolveAppVersionResources API.

Arguments

app_arn: The Amazon Resource Name (ARN) of the AWS Resilience Hub application. The format for this ARN is: arn:partition:resiliencehub:region:account:app/app-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"additionalInfo": Currently, there is no supported additional information for resources.

"appComponents": The list of Application Components that this resource belongs to. If an Application Component is not part of the AWS Resilience Hub application, it will be added.

"awsAccountId": The Amazon Web Services account that owns the physical resource.

"awsRegion": The Amazon Web Services region that owns the physical resource.

"excluded": Indicates if a resource is excluded from an AWS Resilience Hub application. You can exclude only imported resources from an AWS Resilience Hub application.

"logicalResourceId": The logical identifier of the resource.

"physicalResourceId": The physical identifier of the resource.

"resourceName": The name of the resource.

"resourceType": The type of resource.

Main.Resiliencehub.update_resiliency_policy — Method

update_resiliency_policy(policy_arn)
update_resiliency_policy(policy_arn, params::Dict{String,<:Any})

Updates a resiliency policy.

Arguments

policy_arn: The Amazon Resource Name (ARN) of the resiliency policy. The format for this ARN is: arn:partition:resiliencehub:region:account:resiliency-policy/policy-id. For more information about ARNs, see Amazon Resource Names (ARNs) in the AWS General Reference guide.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"dataLocationConstraint": Specifies a high-level geographical location constraint for where your resilience policy data can be stored.

"policy": The type of resiliency policy to be created, including the recovery time objective (RTO) and recovery point objective (RPO) in seconds.

"policyDescription": The description for the policy.

"policyName": The name of the policy

"tier": The tier for this resiliency policy, ranging from the highest severity (MissionCritical) to lowest (NonCritical).