Service Quotas

This page documents function available when using the Service_Quotas module, created with @service Service_Quotas.

Index

Main.Service_Quotas.associate_service_quota_template

Main.Service_Quotas.delete_service_quota_increase_request_from_template

Main.Service_Quotas.disassociate_service_quota_template

Main.Service_Quotas.get_association_for_service_quota_template

Main.Service_Quotas.get_awsdefault_service_quota

Main.Service_Quotas.get_requested_service_quota_change

Main.Service_Quotas.get_service_quota

Main.Service_Quotas.get_service_quota_increase_request_from_template

Main.Service_Quotas.list_awsdefault_service_quotas

Main.Service_Quotas.list_requested_service_quota_change_history

Main.Service_Quotas.list_requested_service_quota_change_history_by_quota

Main.Service_Quotas.list_service_quota_increase_requests_in_template

Main.Service_Quotas.list_service_quotas

Main.Service_Quotas.list_services

Main.Service_Quotas.list_tags_for_resource

Main.Service_Quotas.put_service_quota_increase_request_into_template

Main.Service_Quotas.request_service_quota_increase

Main.Service_Quotas.tag_resource

Main.Service_Quotas.untag_resource

Documentation

Main.Service_Quotas.associate_service_quota_template — Method

associate_service_quota_template()
associate_service_quota_template(params::Dict{String,<:Any})

Associates your quota request template with your organization. When a new account is created in your organization, the quota increase requests in the template are automatically applied to the account. You can add a quota increase request for any adjustable quota to your template.

Main.Service_Quotas.delete_service_quota_increase_request_from_template — Method

delete_service_quota_increase_request_from_template(aws_region, quota_code, service_code)
delete_service_quota_increase_request_from_template(aws_region, quota_code, service_code, params::Dict{String,<:Any})

Deletes the quota increase request for the specified quota from your quota request template.

Arguments

aws_region: The AWS Region.

quota_code: The quota identifier.

service_code: The service identifier.

Main.Service_Quotas.disassociate_service_quota_template — Method

disassociate_service_quota_template()
disassociate_service_quota_template(params::Dict{String,<:Any})

Disables your quota request template. After a template is disabled, the quota increase requests in the template are not applied to new accounts in your organization. Disabling a quota request template does not apply its quota increase requests.

Main.Service_Quotas.get_association_for_service_quota_template — Method

get_association_for_service_quota_template()
get_association_for_service_quota_template(params::Dict{String,<:Any})

Retrieves the status of the association for the quota request template.

Main.Service_Quotas.get_awsdefault_service_quota — Method

get_awsdefault_service_quota(quota_code, service_code)
get_awsdefault_service_quota(quota_code, service_code, params::Dict{String,<:Any})

Retrieves the default value for the specified quota. The default value does not reflect any quota increases.

Arguments

quota_code: The quota identifier.

service_code: The service identifier.

Main.Service_Quotas.get_requested_service_quota_change — Method

get_requested_service_quota_change(request_id)
get_requested_service_quota_change(request_id, params::Dict{String,<:Any})

Retrieves information about the specified quota increase request.

Arguments

request_id: The ID of the quota increase request.

Main.Service_Quotas.get_service_quota — Method

get_service_quota(quota_code, service_code)
get_service_quota(quota_code, service_code, params::Dict{String,<:Any})

Retrieves the applied quota value for the specified quota. For some quotas, only the default values are available. If the applied quota value is not available for a quota, the quota is not retrieved.

Arguments

quota_code: The quota identifier.

service_code: The service identifier.

Main.Service_Quotas.get_service_quota_increase_request_from_template — Method

get_service_quota_increase_request_from_template(aws_region, quota_code, service_code)
get_service_quota_increase_request_from_template(aws_region, quota_code, service_code, params::Dict{String,<:Any})

Retrieves information about the specified quota increase request in your quota request template.

Arguments

aws_region: The AWS Region.

quota_code: The quota identifier.

service_code: The service identifier.

Main.Service_Quotas.list_awsdefault_service_quotas — Method

list_awsdefault_service_quotas(service_code)
list_awsdefault_service_quotas(service_code, params::Dict{String,<:Any})

Lists the default values for the quotas for the specified AWS service. A default value does not reflect any quota increases.

Arguments

service_code: The service identifier.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return with a single call. To retrieve the remaining results, if any, make another call with the token returned from this call.

"NextToken": The token for the next page of results.

Main.Service_Quotas.list_requested_service_quota_change_history — Method

list_requested_service_quota_change_history()
list_requested_service_quota_change_history(params::Dict{String,<:Any})

Retrieves the quota increase requests for the specified service.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return with a single call. To retrieve the remaining results, if any, make another call with the token returned from this call.

"NextToken": The token for the next page of results.

"ServiceCode": The service identifier.

"Status": The status of the quota increase request.

Main.Service_Quotas.list_requested_service_quota_change_history_by_quota — Method

list_requested_service_quota_change_history_by_quota(quota_code, service_code)
list_requested_service_quota_change_history_by_quota(quota_code, service_code, params::Dict{String,<:Any})

Retrieves the quota increase requests for the specified quota.

Arguments

quota_code: The quota identifier.

service_code: The service identifier.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return with a single call. To retrieve the remaining results, if any, make another call with the token returned from this call.

"NextToken": The token for the next page of results.

"Status": The status value of the quota increase request.

Main.Service_Quotas.list_service_quota_increase_requests_in_template — Method

list_service_quota_increase_requests_in_template()
list_service_quota_increase_requests_in_template(params::Dict{String,<:Any})

Lists the quota increase requests in the specified quota request template.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"AwsRegion": The AWS Region.

"MaxResults": The maximum number of results to return with a single call. To retrieve the remaining results, if any, make another call with the token returned from this call.

"NextToken": The token for the next page of results.

"ServiceCode": The service identifier.

Main.Service_Quotas.list_service_quotas — Method

list_service_quotas(service_code)
list_service_quotas(service_code, params::Dict{String,<:Any})

Lists the applied quota values for the specified AWS service. For some quotas, only the default values are available. If the applied quota value is not available for a quota, the quota is not retrieved.

Arguments

service_code: The service identifier.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return with a single call. To retrieve the remaining results, if any, make another call with the token returned from this call.

"NextToken": The token for the next page of results.

Main.Service_Quotas.list_services — Method

list_services()
list_services(params::Dict{String,<:Any})

Lists the names and codes for the services integrated with Service Quotas.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": The maximum number of results to return with a single call. To retrieve the remaining results, if any, make another call with the token returned from this call.

"NextToken": The token for the next page of results.

Main.Service_Quotas.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Returns a list of the tags assigned to the specified applied quota.

Arguments

resource_arn: The Amazon Resource Name (ARN) for the applied quota for which you want to list tags. You can get this information by using the Service Quotas console, or by listing the quotas using the list-service-quotas AWS CLI command or the ListServiceQuotas AWS API operation.

Main.Service_Quotas.put_service_quota_increase_request_into_template — Method

put_service_quota_increase_request_into_template(aws_region, desired_value, quota_code, service_code)
put_service_quota_increase_request_into_template(aws_region, desired_value, quota_code, service_code, params::Dict{String,<:Any})

Adds a quota increase request to your quota request template.

Arguments

aws_region: The AWS Region.

desired_value: The new, increased value for the quota.

quota_code: The quota identifier.

service_code: The service identifier.

Main.Service_Quotas.request_service_quota_increase — Method

request_service_quota_increase(desired_value, quota_code, service_code)
request_service_quota_increase(desired_value, quota_code, service_code, params::Dict{String,<:Any})

Submits a quota increase request for the specified quota.

Arguments

desired_value: The new, increased value for the quota.

quota_code: The quota identifier.

service_code: The service identifier.

Main.Service_Quotas.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Adds tags to the specified applied quota. You can include one or more tags to add to the quota.

Arguments

resource_arn: The Amazon Resource Name (ARN) for the applied quota. You can get this information by using the Service Quotas console, or by listing the quotas using the list-service-quotas AWS CLI command or the ListServiceQuotas AWS API operation.

tags: The tags that you want to add to the resource.

Main.Service_Quotas.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes tags from the specified applied quota. You can specify one or more tags to remove.

Arguments

resource_arn: The Amazon Resource Name (ARN) for the applied quota that you want to untag. You can get this information by using the Service Quotas console, or by listing the quotas using the list-service-quotas AWS CLI command or the ListServiceQuotas AWS API operation.

tag_keys: The keys of the tags that you want to remove from the resource.