Sfn

This page documents function available when using the Sfn module, created with @service Sfn.

Index

Main.Sfn.create_activity

Main.Sfn.create_state_machine

Main.Sfn.delete_activity

Main.Sfn.delete_state_machine

Main.Sfn.describe_activity

Main.Sfn.describe_execution

Main.Sfn.describe_state_machine

Main.Sfn.describe_state_machine_for_execution

Main.Sfn.get_activity_task

Main.Sfn.get_execution_history

Main.Sfn.list_activities

Main.Sfn.list_executions

Main.Sfn.list_state_machines

Main.Sfn.list_tags_for_resource

Main.Sfn.send_task_failure

Main.Sfn.send_task_heartbeat

Main.Sfn.send_task_success

Main.Sfn.start_execution

Main.Sfn.start_sync_execution

Main.Sfn.stop_execution

Main.Sfn.tag_resource

Main.Sfn.untag_resource

Main.Sfn.update_state_machine

Documentation

Main.Sfn.create_activity — Method

create_activity(name)
create_activity(name, params::Dict{String,<:Any})

Creates an activity. An activity is a task that you write in any programming language and host on any machine that has access to AWS Step Functions. Activities must poll Step Functions using the GetActivityTask API action and respond using SendTask* API actions. This function lets Step Functions know the existence of your activity and returns an identifier for use in a state machine and when polling from the activity. This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes. CreateActivity is an idempotent API. Subsequent requests won’t create a duplicate resource if it was already created. CreateActivity's idempotency check is based on the activity name. If a following request has different tags values, Step Functions will ignore these differences and treat it as an idempotent request of the previous. In this case, tags will not be updated, even if they are different.

Arguments

name: The name of the activity to create. This name must be unique for your AWS account and region for 90 days. For more information, see Limits Related to State Machine Executions in the AWS Step Functions Developer Guide. A name must not contain: white space brackets &lt; &gt; { } [ ] wildcard characters ? * special characters " # % ^ | ~ ` &amp; , ; : / control characters (U+0000-001F, U+007F-009F) To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"tags": The list of tags to add to a resource. An array of key-value pairs. For more information, see Using Cost Allocation Tags in the AWS Billing and Cost Management User Guide, and Controlling Access Using IAM Tags. Tags may only contain Unicode letters, digits, white space, or these symbols: _ . : / = + - @.

Main.Sfn.create_state_machine — Method

create_state_machine(definition, name, role_arn)
create_state_machine(definition, name, role_arn, params::Dict{String,<:Any})

Creates a state machine. A state machine consists of a collection of states that can do work (Task states), determine to which states to transition next (Choice states), stop an execution with an error (Fail states), and so on. State machines are specified using a JSON-based, structured language. For more information, see Amazon States Language in the AWS Step Functions User Guide. This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes. CreateStateMachine is an idempotent API. Subsequent requests won’t create a duplicate resource if it was already created. CreateStateMachine's idempotency check is based on the state machine name, definition, type, LoggingConfiguration and TracingConfiguration. If a following request has a different roleArn or tags, Step Functions will ignore these differences and treat it as an idempotent request of the previous. In this case, roleArn and tags will not be updated, even if they are different.

Arguments

definition: The Amazon States Language definition of the state machine. See Amazon States Language.

name: The name of the state machine. A name must not contain: white space brackets &lt; &gt; { } [ ] wildcard characters ? * special characters " # % ^ | ~ ` &amp; , ; : / control characters (U+0000-001F, U+007F-009F) To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

role_arn: The Amazon Resource Name (ARN) of the IAM role to use for this state machine.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"loggingConfiguration": Defines what execution history events are logged and where they are logged. By default, the level is set to OFF. For more information see Log Levels in the AWS Step Functions User Guide.

"tags": Tags to be added when creating a state machine. An array of key-value pairs. For more information, see Using Cost Allocation Tags in the AWS Billing and Cost Management User Guide, and Controlling Access Using IAM Tags. Tags may only contain Unicode letters, digits, white space, or these symbols: _ . : / = + - @.

"tracingConfiguration": Selects whether AWS X-Ray tracing is enabled.

"type": Determines whether a Standard or Express state machine is created. The default is STANDARD. You cannot update the type of a state machine once it has been created.

Main.Sfn.delete_activity — Method

delete_activity(activity_arn)
delete_activity(activity_arn, params::Dict{String,<:Any})

Deletes an activity.

Arguments

activity_arn: The Amazon Resource Name (ARN) of the activity to delete.

Main.Sfn.delete_state_machine — Method

delete_state_machine(state_machine_arn)
delete_state_machine(state_machine_arn, params::Dict{String,<:Any})

Deletes a state machine. This is an asynchronous operation: It sets the state machine's status to DELETING and begins the deletion process. For EXPRESSstate machines, the deletion will happen eventually (usually less than a minute). Running executions may emit logs after DeleteStateMachine API is called.

Arguments

state_machine_arn: The Amazon Resource Name (ARN) of the state machine to delete.

Main.Sfn.describe_activity — Method

describe_activity(activity_arn)
describe_activity(activity_arn, params::Dict{String,<:Any})

Describes an activity. This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes.

Arguments

activity_arn: The Amazon Resource Name (ARN) of the activity to describe.

Main.Sfn.describe_execution — Method

describe_execution(execution_arn)
describe_execution(execution_arn, params::Dict{String,<:Any})

Describes an execution. This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes. This API action is not supported by EXPRESS state machines.

Arguments

execution_arn: The Amazon Resource Name (ARN) of the execution to describe.

Main.Sfn.describe_state_machine — Method

describe_state_machine(state_machine_arn)
describe_state_machine(state_machine_arn, params::Dict{String,<:Any})

Describes a state machine. This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes.

Arguments

state_machine_arn: The Amazon Resource Name (ARN) of the state machine to describe.

Main.Sfn.describe_state_machine_for_execution — Method

describe_state_machine_for_execution(execution_arn)
describe_state_machine_for_execution(execution_arn, params::Dict{String,<:Any})

Describes the state machine associated with a specific execution. This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes. This API action is not supported by EXPRESS state machines.

Arguments

execution_arn: The Amazon Resource Name (ARN) of the execution you want state machine information for.

Main.Sfn.get_activity_task — Method

get_activity_task(activity_arn)
get_activity_task(activity_arn, params::Dict{String,<:Any})

Used by workers to retrieve a task (with the specified activity ARN) which has been scheduled for execution by a running state machine. This initiates a long poll, where the service holds the HTTP connection open and responds as soon as a task becomes available (i.e. an execution of a task of this type is needed.) The maximum time the service holds on to the request before responding is 60 seconds. If no task is available within 60 seconds, the poll returns a taskToken with a null string. Workers should set their client side socket timeout to at least 65 seconds (5 seconds higher than the maximum time the service may hold the poll request). Polling with GetActivityTask can cause latency in some implementations. See Avoid Latency When Polling for Activity Tasks in the Step Functions Developer Guide.

Arguments

activity_arn: The Amazon Resource Name (ARN) of the activity to retrieve tasks from (assigned when you create the task using CreateActivity.)

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"workerName": You can provide an arbitrary name in order to identify the worker that the task is assigned to. This name is used when it is logged in the execution history.

Main.Sfn.get_execution_history — Method

get_execution_history(execution_arn)
get_execution_history(execution_arn, params::Dict{String,<:Any})

Returns the history of the specified execution as a list of events. By default, the results are returned in ascending order of the timeStamp of the events. Use the reverseOrder parameter to get the latest events first. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error. This API action is not supported by EXPRESS state machines.

Arguments

execution_arn: The Amazon Resource Name (ARN) of the execution.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"includeExecutionData": You can select whether execution data (input or output of a history event) is returned. The default is true.

"maxResults": The maximum number of results that are returned per call. You can use nextToken to obtain further pages of results. The default is 100 and the maximum allowed page size is 1000. A value of 0 uses the default. This is only an upper limit. The actual number of results returned per call might be fewer than the specified maximum.

"nextToken": If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

"reverseOrder": Lists events in descending order of their timeStamp.

Main.Sfn.list_activities — Method

list_activities()
list_activities(params::Dict{String,<:Any})

Lists the existing activities. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error. This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results that are returned per call. You can use nextToken to obtain further pages of results. The default is 100 and the maximum allowed page size is 1000. A value of 0 uses the default. This is only an upper limit. The actual number of results returned per call might be fewer than the specified maximum.

"nextToken": If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

Main.Sfn.list_executions — Method

list_executions(state_machine_arn)
list_executions(state_machine_arn, params::Dict{String,<:Any})

Lists the executions of a state machine that meet the filtering criteria. Results are sorted by time, with the most recent execution first. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error. This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes. This API action is not supported by EXPRESS state machines.

Arguments

state_machine_arn: The Amazon Resource Name (ARN) of the state machine whose executions is listed.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results that are returned per call. You can use nextToken to obtain further pages of results. The default is 100 and the maximum allowed page size is 1000. A value of 0 uses the default. This is only an upper limit. The actual number of results returned per call might be fewer than the specified maximum.

"nextToken": If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

"statusFilter": If specified, only list the executions whose current execution status matches the given filter.

Main.Sfn.list_state_machines — Method

list_state_machines()
list_state_machines(params::Dict{String,<:Any})

Lists the existing state machines. If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error. This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"maxResults": The maximum number of results that are returned per call. You can use nextToken to obtain further pages of results. The default is 100 and the maximum allowed page size is 1000. A value of 0 uses the default. This is only an upper limit. The actual number of results returned per call might be fewer than the specified maximum.

"nextToken": If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

Main.Sfn.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

List tags for a given resource. Tags may only contain Unicode letters, digits, white space, or these symbols: _ . : / = + - @.

Arguments

resource_arn: The Amazon Resource Name (ARN) for the Step Functions state machine or activity.

Main.Sfn.send_task_failure — Method

send_task_failure(task_token)
send_task_failure(task_token, params::Dict{String,<:Any})

Used by activity workers and task states using the callback pattern to report that the task identified by the taskToken failed.

Arguments

task_token: The token that represents this task. Task tokens are generated by Step Functions when tasks are assigned to a worker, or in the context object when a workflow enters a task state. See GetActivityTaskOutputtaskToken.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"cause": A more detailed explanation of the cause of the failure.

"error": The error code of the failure.

Main.Sfn.send_task_heartbeat — Method

send_task_heartbeat(task_token)
send_task_heartbeat(task_token, params::Dict{String,<:Any})

Used by activity workers and task states using the callback pattern to report to Step Functions that the task represented by the specified taskToken is still making progress. This action resets the Heartbeat clock. The Heartbeat threshold is specified in the state machine's Amazon States Language definition (HeartbeatSeconds). This action does not in itself create an event in the execution history. However, if the task times out, the execution history contains an ActivityTimedOut entry for activities, or a TaskTimedOut entry for for tasks using the job run or callback pattern. The Timeout of a task, defined in the state machine's Amazon States Language definition, is its maximum allowed duration, regardless of the number of SendTaskHeartbeat requests received. Use HeartbeatSeconds to configure the timeout interval for heartbeats.

Arguments

task_token: The token that represents this task. Task tokens are generated by Step Functions when tasks are assigned to a worker, or in the context object when a workflow enters a task state. See GetActivityTaskOutputtaskToken.

Main.Sfn.send_task_success — Method

send_task_success(output, task_token)
send_task_success(output, task_token, params::Dict{String,<:Any})

Used by activity workers and task states using the callback pattern to report that the task identified by the taskToken completed successfully.

Arguments

output: The JSON output of the task. Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

task_token: The token that represents this task. Task tokens are generated by Step Functions when tasks are assigned to a worker, or in the context object when a workflow enters a task state. See GetActivityTaskOutputtaskToken.

Main.Sfn.start_execution — Method

start_execution(state_machine_arn)
start_execution(state_machine_arn, params::Dict{String,<:Any})

Starts a state machine execution. StartExecution is idempotent. If StartExecution is called with the same name and input as a running execution, the call will succeed and return the same response as the original request. If the execution is closed or if the input is different, it will return a 400 ExecutionAlreadyExists error. Names can be reused after 90 days.

Arguments

state_machine_arn: The Amazon Resource Name (ARN) of the state machine to execute.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"input": The string that contains the JSON input data for the execution, for example: "input": "{"first_name" : "test"}" If you don't include any JSON input data, you still must include the two braces, for example: "input": "{}" Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

"name": The name of the execution. This name must be unique for your AWS account, region, and state machine for 90 days. For more information, see Limits Related to State Machine Executions in the AWS Step Functions Developer Guide. A name must not contain: white space brackets &lt; &gt; { } [ ] wildcard characters ? * special characters " # % ^ | ~ ` &amp; , ; : / control characters (U+0000-001F, U+007F-009F) To enable logging with CloudWatch Logs, the name should only contain 0-9, A-Z, a-z, - and _.

"traceHeader": Passes the AWS X-Ray trace header. The trace header can also be passed in the request payload.

Main.Sfn.start_sync_execution — Method

start_sync_execution(state_machine_arn)
start_sync_execution(state_machine_arn, params::Dict{String,<:Any})

Starts a Synchronous Express state machine execution.

Arguments

state_machine_arn: The Amazon Resource Name (ARN) of the state machine to execute.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"input": The string that contains the JSON input data for the execution, for example: "input": "{"first_name" : "test"}" If you don't include any JSON input data, you still must include the two braces, for example: "input": "{}" Length constraints apply to the payload size, and are expressed as bytes in UTF-8 encoding.

"name": The name of the execution.

"traceHeader": Passes the AWS X-Ray trace header. The trace header can also be passed in the request payload.

Main.Sfn.stop_execution — Method

stop_execution(execution_arn)
stop_execution(execution_arn, params::Dict{String,<:Any})

Stops an execution. This API action is not supported by EXPRESS state machines.

Arguments

execution_arn: The Amazon Resource Name (ARN) of the execution to stop.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"cause": A more detailed explanation of the cause of the failure.

"error": The error code of the failure.

Main.Sfn.tag_resource — Method

tag_resource(resource_arn, tags)
tag_resource(resource_arn, tags, params::Dict{String,<:Any})

Add a tag to a Step Functions resource. An array of key-value pairs. For more information, see Using Cost Allocation Tags in the AWS Billing and Cost Management User Guide, and Controlling Access Using IAM Tags. Tags may only contain Unicode letters, digits, white space, or these symbols: _ . : / = + - @.

Arguments

resource_arn: The Amazon Resource Name (ARN) for the Step Functions state machine or activity.

tags: The list of tags to add to a resource. Tags may only contain Unicode letters, digits, white space, or these symbols: _ . : / = + - @.

Main.Sfn.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Remove a tag from a Step Functions resource

Arguments

resource_arn: The Amazon Resource Name (ARN) for the Step Functions state machine or activity.

tag_keys: The list of tags to remove from the resource.

Main.Sfn.update_state_machine — Method

update_state_machine(state_machine_arn)
update_state_machine(state_machine_arn, params::Dict{String,<:Any})

Updates an existing state machine by modifying its definition, roleArn, or loggingConfiguration. Running executions will continue to use the previous definition and roleArn. You must include at least one of definition or roleArn or you will receive a MissingRequiredParameter error. All StartExecution calls within a few seconds will use the updated definition and roleArn. Executions started immediately after calling UpdateStateMachine may use the previous state machine definition and roleArn.

Arguments

state_machine_arn: The Amazon Resource Name (ARN) of the state machine.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"definition": The Amazon States Language definition of the state machine. See Amazon States Language.

"loggingConfiguration": The LoggingConfiguration data type is used to set CloudWatch Logs options.

"roleArn": The Amazon Resource Name (ARN) of the IAM role of the state machine.

"tracingConfiguration": Selects whether AWS X-Ray tracing is enabled.