Synthetics

This page documents function available when using the Synthetics module, created with @service Synthetics.

Index

Main.Synthetics.create_canary

Main.Synthetics.delete_canary

Main.Synthetics.describe_canaries

Main.Synthetics.describe_canaries_last_run

Main.Synthetics.describe_runtime_versions

Main.Synthetics.get_canary

Main.Synthetics.get_canary_runs

Main.Synthetics.list_tags_for_resource

Main.Synthetics.start_canary

Main.Synthetics.stop_canary

Main.Synthetics.tag_resource

Main.Synthetics.untag_resource

Main.Synthetics.update_canary

Documentation

Main.Synthetics.create_canary — Method

create_canary(artifact_s3_location, code, execution_role_arn, name, runtime_version, schedule)
create_canary(artifact_s3_location, code, execution_role_arn, name, runtime_version, schedule, params::Dict{String,<:Any})

Creates a canary. Canaries are scripts that monitor your endpoints and APIs from the outside-in. Canaries help you check the availability and latency of your web services and troubleshoot anomalies by investigating load time data, screenshots of the UI, logs, and metrics. You can set up a canary to run continuously or just once. Do not use CreateCanary to modify an existing canary. Use UpdateCanary instead. To create canaries, you must have the CloudWatchSyntheticsFullAccess policy. If you are creating a new IAM role for the canary, you also need the the iam:CreateRole, iam:CreatePolicy and iam:AttachRolePolicy permissions. For more information, see Necessary Roles and Permissions. Do not include secrets or proprietary information in your canary names. The canary name makes up part of the Amazon Resource Name (ARN) for the canary, and the ARN is included in outbound calls over the internet. For more information, see Security Considerations for Synthetics Canaries.

Arguments

artifact_s3_location: The location in Amazon S3 where Synthetics stores artifacts from the test runs of this canary. Artifacts include the log file, screenshots, and HAR files.

code: A structure that includes the entry point from which the canary should start running your script. If the script is stored in an S3 bucket, the bucket name, key, and version are also included.

execution_role_arn: The ARN of the IAM role to be used to run the canary. This role must already exist, and must include lambda.amazonaws.com as a principal in the trust policy. The role must also have the following permissions: s3:PutObject s3:GetBucketLocation s3:ListAllMyBuckets cloudwatch:PutMetricData logs:CreateLogGroup logs:CreateLogStream logs:PutLogEvents

name: The name for this canary. Be sure to give it a descriptive name that distinguishes it from other canaries in your account. Do not include secrets or proprietary information in your canary names. The canary name makes up part of the canary ARN, and the ARN is included in outbound calls over the internet. For more information, see Security Considerations for Synthetics Canaries.

runtime_version: Specifies the runtime version to use for the canary. For a list of valid runtime versions and more information about runtime versions, see Canary Runtime Versions.

schedule: A structure that contains information about how often the canary is to run and when these test runs are to stop.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"FailureRetentionPeriodInDays": The number of days to retain data about failed runs of this canary. If you omit this field, the default of 31 days is used. The valid range is 1 to 455 days.

"RunConfig": A structure that contains the configuration for individual canary runs, such as timeout value.

"SuccessRetentionPeriodInDays": The number of days to retain data about successful runs of this canary. If you omit this field, the default of 31 days is used. The valid range is 1 to 455 days.

"Tags": A list of key-value pairs to associate with the canary. You can associate as many as 50 tags with a canary. Tags can help you organize and categorize your resources. You can also use them to scope user permissions, by granting a user permission to access or change only the resources that have certain tag values.

"VpcConfig": If this canary is to test an endpoint in a VPC, this structure contains information about the subnet and security groups of the VPC endpoint. For more information, see Running a Canary in a VPC.

Main.Synthetics.delete_canary — Method

delete_canary(name)
delete_canary(name, params::Dict{String,<:Any})

Permanently deletes the specified canary. When you delete a canary, resources used and created by the canary are not automatically deleted. After you delete a canary that you do not intend to use again, you should also delete the following: The Lambda functions and layers used by this canary. These have the prefix cwsyn-MyCanaryName . The CloudWatch alarms created for this canary. These alarms have a name of Synthetics-SharpDrop-Alarm-MyCanaryName . Amazon S3 objects and buckets, such as the canary's artifact location. IAM roles created for the canary. If they were created in the console, these roles have the name role/service-role/CloudWatchSyntheticsRole-MyCanaryName . CloudWatch Logs log groups created for the canary. These logs groups have the name /aws/lambda/cwsyn-MyCanaryName . Before you delete a canary, you might want to use GetCanary to display the information about this canary. Make note of the information returned by this operation so that you can delete these resources after you delete the canary.

Arguments

name: The name of the canary that you want to delete. To find the names of your canaries, use DescribeCanaries.

Main.Synthetics.describe_canaries — Method

describe_canaries()
describe_canaries(params::Dict{String,<:Any})

This operation returns a list of the canaries in your account, along with full details about each canary. This operation does not have resource-level authorization, so if a user is able to use DescribeCanaries, the user can see all of the canaries in the account. A deny policy can only be used to restrict access to all canaries. It cannot be used on specific resources.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": Specify this parameter to limit how many canaries are returned each time you use the DescribeCanaries operation. If you omit this parameter, the default of 100 is used.

"NextToken": A token that indicates that there is more data available. You can use this token in a subsequent operation to retrieve the next set of results.

Main.Synthetics.describe_canaries_last_run — Method

describe_canaries_last_run()
describe_canaries_last_run(params::Dict{String,<:Any})

Use this operation to see information from the most recent run of each canary that you have created.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": Specify this parameter to limit how many runs are returned each time you use the DescribeLastRun operation. If you omit this parameter, the default of 100 is used.

"NextToken": A token that indicates that there is more data available. You can use this token in a subsequent DescribeCanaries operation to retrieve the next set of results.

Main.Synthetics.describe_runtime_versions — Method

describe_runtime_versions()
describe_runtime_versions(params::Dict{String,<:Any})

Returns a list of Synthetics canary runtime versions. For more information, see Canary Runtime Versions.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": Specify this parameter to limit how many runs are returned each time you use the DescribeRuntimeVersions operation. If you omit this parameter, the default of 100 is used.

"NextToken": A token that indicates that there is more data available. You can use this token in a subsequent DescribeRuntimeVersions operation to retrieve the next set of results.

Main.Synthetics.get_canary — Method

get_canary(name)
get_canary(name, params::Dict{String,<:Any})

Retrieves complete information about one canary. You must specify the name of the canary that you want. To get a list of canaries and their names, use DescribeCanaries.

Arguments

name: The name of the canary that you want details for.

Main.Synthetics.get_canary_runs — Method

get_canary_runs(name)
get_canary_runs(name, params::Dict{String,<:Any})

Retrieves a list of runs for a specified canary.

Arguments

name: The name of the canary that you want to see runs for.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"MaxResults": Specify this parameter to limit how many runs are returned each time you use the GetCanaryRuns operation. If you omit this parameter, the default of 100 is used.

"NextToken": A token that indicates that there is more data available. You can use this token in a subsequent GetCanaryRuns operation to retrieve the next set of results.

Main.Synthetics.list_tags_for_resource — Method

list_tags_for_resource(resource_arn)
list_tags_for_resource(resource_arn, params::Dict{String,<:Any})

Displays the tags associated with a canary.

Arguments

resource_arn: The ARN of the canary that you want to view tags for. The ARN format of a canary is arn:aws:synthetics:Region:account-id:canary:canary-name .

Main.Synthetics.start_canary — Method

start_canary(name)
start_canary(name, params::Dict{String,<:Any})

Use this operation to run a canary that has already been created. The frequency of the canary runs is determined by the value of the canary's Schedule. To see a canary's schedule, use GetCanary.

Arguments

name: The name of the canary that you want to run. To find canary names, use DescribeCanaries.

Main.Synthetics.stop_canary — Method

stop_canary(name)
stop_canary(name, params::Dict{String,<:Any})

Stops the canary to prevent all future runs. If the canary is currently running, Synthetics stops waiting for the current run of the specified canary to complete. The run that is in progress completes on its own, publishes metrics, and uploads artifacts, but it is not recorded in Synthetics as a completed run. You can use StartCanary to start it running again with the canary’s current schedule at any point in the future.

Arguments

name: The name of the canary that you want to stop. To find the names of your canaries, use DescribeCanaries.

Main.Synthetics.tag_resource — Method

tag_resource(tags, resource_arn)
tag_resource(tags, resource_arn, params::Dict{String,<:Any})

Assigns one or more tags (key-value pairs) to the specified canary. Tags can help you organize and categorize your resources. You can also use them to scope user permissions, by granting a user permission to access or change only resources with certain tag values. Tags don't have any semantic meaning to AWS and are interpreted strictly as strings of characters. You can use the TagResource action with a canary that already has tags. If you specify a new tag key for the alarm, this tag is appended to the list of tags associated with the alarm. If you specify a tag key that is already associated with the alarm, the new tag value that you specify replaces the previous value for that tag. You can associate as many as 50 tags with a canary.

Arguments

tags: The list of key-value pairs to associate with the canary.

resource_arn: The ARN of the canary that you're adding tags to. The ARN format of a canary is arn:aws:synthetics:Region:account-id:canary:canary-name .

Main.Synthetics.untag_resource — Method

untag_resource(resource_arn, tag_keys)
untag_resource(resource_arn, tag_keys, params::Dict{String,<:Any})

Removes one or more tags from the specified canary.

Arguments

resource_arn: The ARN of the canary that you're removing tags from. The ARN format of a canary is arn:aws:synthetics:Region:account-id:canary:canary-name .

tag_keys: The list of tag keys to remove from the resource.

Main.Synthetics.update_canary — Method

update_canary(name)
update_canary(name, params::Dict{String,<:Any})

Use this operation to change the settings of a canary that has already been created. You can't use this operation to update the tags of an existing canary. To change the tags of an existing canary, use TagResource.

Arguments

name: The name of the canary that you want to update. To find the names of your canaries, use DescribeCanaries. You cannot change the name of a canary that has already been created.

Optional Parameters

Optional parameters can be passed as a params::Dict{String,<:Any}. Valid keys are:

"Code": A structure that includes the entry point from which the canary should start running your script. If the script is stored in an S3 bucket, the bucket name, key, and version are also included.

"ExecutionRoleArn": The ARN of the IAM role to be used to run the canary. This role must already exist, and must include lambda.amazonaws.com as a principal in the trust policy. The role must also have the following permissions: s3:PutObject s3:GetBucketLocation s3:ListAllMyBuckets cloudwatch:PutMetricData logs:CreateLogGroup logs:CreateLogStream logs:CreateLogStream

"FailureRetentionPeriodInDays": The number of days to retain data about failed runs of this canary.

"RunConfig": A structure that contains the timeout value that is used for each individual run of the canary.

"RuntimeVersion": Specifies the runtime version to use for the canary. For a list of valid runtime versions and for more information about runtime versions, see Canary Runtime Versions.

"Schedule": A structure that contains information about how often the canary is to run, and when these runs are to stop.

"SuccessRetentionPeriodInDays": The number of days to retain data about successful runs of this canary.

"VpcConfig": If this canary is to test an endpoint in a VPC, this structure contains information about the subnet and security groups of the VPC endpoint. For more information, see Running a Canary in a VPC.