Workspaces

This page documents function available when using the Workspaces module, created with @service Workspaces.

Index

Main.Workspaces.associate_connection_alias

Main.Workspaces.associate_ip_groups

Main.Workspaces.authorize_ip_rules

Main.Workspaces.copy_workspace_image

Main.Workspaces.create_connection_alias

Main.Workspaces.create_ip_group

Main.Workspaces.create_tags

Main.Workspaces.create_workspaces

Main.Workspaces.delete_connection_alias

Main.Workspaces.delete_ip_group

Main.Workspaces.delete_tags

Main.Workspaces.delete_workspace_image

Main.Workspaces.deregister_workspace_directory

Main.Workspaces.describe_account

Main.Workspaces.describe_account_modifications

Main.Workspaces.describe_client_properties

Main.Workspaces.describe_connection_alias_permissions

Main.Workspaces.describe_connection_aliases

Main.Workspaces.describe_ip_groups

Main.Workspaces.describe_tags

Main.Workspaces.describe_workspace_bundles

Main.Workspaces.describe_workspace_directories

Main.Workspaces.describe_workspace_image_permissions

Main.Workspaces.describe_workspace_images

Main.Workspaces.describe_workspace_snapshots

Main.Workspaces.describe_workspaces

Main.Workspaces.describe_workspaces_connection_status

Main.Workspaces.disassociate_connection_alias

Main.Workspaces.disassociate_ip_groups

Main.Workspaces.import_workspace_image

Main.Workspaces.list_available_management_cidr_ranges

Main.Workspaces.migrate_workspace

Main.Workspaces.modify_account

Main.Workspaces.modify_client_properties

Main.Workspaces.modify_selfservice_permissions

Main.Workspaces.modify_workspace_access_properties

Main.Workspaces.modify_workspace_creation_properties

Main.Workspaces.modify_workspace_properties

Main.Workspaces.modify_workspace_state

Main.Workspaces.reboot_workspaces

Main.Workspaces.rebuild_workspaces

Main.Workspaces.register_workspace_directory

Main.Workspaces.restore_workspace

Main.Workspaces.revoke_ip_rules

Main.Workspaces.start_workspaces

Main.Workspaces.stop_workspaces

Main.Workspaces.terminate_workspaces

Main.Workspaces.update_connection_alias_permission

Main.Workspaces.update_rules_of_ip_group

Main.Workspaces.update_workspace_image_permission

Documentation

Main.Workspaces.associate_connection_alias — Method

AssociateConnectionAlias()

Associates the specified connection alias with the specified directory to enable cross-Region redirection. For more information, see Cross-Region Redirection for Amazon WorkSpaces. Before performing this operation, call DescribeConnectionAliases to make sure that the current state of the connection alias is CREATED.

Required Parameters

AliasId: The identifier of the connection alias.

ResourceId: The identifier of the directory to associate the connection alias with.

Main.Workspaces.associate_ip_groups — Method

AssociateIpGroups()

Associates the specified IP access control group with the specified directory.

Required Parameters

DirectoryId: The identifier of the directory.

GroupIds: The identifiers of one or more IP access control groups.

Main.Workspaces.authorize_ip_rules — Method

AuthorizeIpRules()

Adds one or more rules to the specified IP access control group. This action gives users permission to access their WorkSpaces from the CIDR address ranges specified in the rules.

Required Parameters

GroupId: The identifier of the group.

UserRules: The rules to add to the group.

Main.Workspaces.copy_workspace_image — Method

CopyWorkspaceImage()

Copies the specified image from the specified Region to the current Region. For more information about copying images, see Copy a Custom WorkSpaces Image. In the China (Ningxia) Region, you can copy images only within the same Region. In the AWS GovCloud (US-West) Region, to copy images to and from other AWS Regions, contact AWS Support. Before copying a shared image, be sure to verify that it has been shared from the correct AWS account. To determine if an image has been shared and to see the AWS account ID that owns an image, use the DescribeWorkSpaceImages and DescribeWorkspaceImagePermissions API operations.

Required Parameters

Name: The name of the image.

SourceImageId: The identifier of the source image.

SourceRegion: The identifier of the source Region.

Optional Parameters

Description: A description of the image.

Tags: The tags for the image.

Main.Workspaces.create_connection_alias — Method

CreateConnectionAlias()

Creates the specified connection alias for use with cross-Region redirection. For more information, see Cross-Region Redirection for Amazon WorkSpaces.

Required Parameters

ConnectionString: A connection string in the form of a fully qualified domain name (FQDN), such as www.example.com. After you create a connection string, it is always associated to your AWS account. You cannot recreate the same connection string with a different account, even if you delete all instances of it from the original account. The connection string is globally reserved for your account.

Optional Parameters

Tags: The tags to associate with the connection alias.

Main.Workspaces.create_ip_group — Method

CreateIpGroup()

Creates an IP access control group. An IP access control group provides you with the ability to control the IP addresses from which users are allowed to access their WorkSpaces. To specify the CIDR address ranges, add rules to your IP access control group and then associate the group with your directory. You can add rules when you create the group or at any time using AuthorizeIpRules. There is a default IP access control group associated with your directory. If you don't associate an IP access control group with your directory, the default group is used. The default group includes a default rule that allows users to access their WorkSpaces from anywhere. You cannot modify the default IP access control group for your directory.

Required Parameters

GroupName: The name of the group.

Optional Parameters

GroupDesc: The description of the group.

Tags: The tags. Each WorkSpaces resource can have a maximum of 50 tags.

UserRules: The rules to add to the group.

Main.Workspaces.create_tags — Method

CreateTags()

Creates the specified tags for the specified WorkSpaces resource.

Required Parameters

ResourceId: The identifier of the WorkSpaces resource. The supported resource types are WorkSpaces, registered directories, images, custom bundles, IP access control groups, and connection aliases.

Tags: The tags. Each WorkSpaces resource can have a maximum of 50 tags.

Main.Workspaces.create_workspaces — Method

CreateWorkspaces()

Creates one or more WorkSpaces. This operation is asynchronous and returns before the WorkSpaces are created.

Required Parameters

Workspaces: The WorkSpaces to create. You can specify up to 25 WorkSpaces.

Main.Workspaces.delete_connection_alias — Method

DeleteConnectionAlias()

Deletes the specified connection alias. For more information, see Cross-Region Redirection for Amazon WorkSpaces. If you will no longer be using a fully qualified domain name (FQDN) as the registration code for your WorkSpaces users, you must take certain precautions to prevent potential security issues. For more information, see Security Considerations if You Stop Using Cross-Region Redirection. To delete a connection alias that has been shared, the shared account must first disassociate the connection alias from any directories it has been associated with. Then you must unshare the connection alias from the account it has been shared with. You can delete a connection alias only after it is no longer shared with any accounts or associated with any directories.

Required Parameters

AliasId: The identifier of the connection alias to delete.

Main.Workspaces.delete_ip_group — Method

DeleteIpGroup()

Deletes the specified IP access control group. You cannot delete an IP access control group that is associated with a directory.

Required Parameters

GroupId: The identifier of the IP access control group.

Main.Workspaces.delete_tags — Method

DeleteTags()

Deletes the specified tags from the specified WorkSpaces resource.

Required Parameters

ResourceId: The identifier of the WorkSpaces resource. The supported resource types are WorkSpaces, registered directories, images, custom bundles, IP access control groups, and connection aliases.

TagKeys: The tag keys.

Main.Workspaces.delete_workspace_image — Method

DeleteWorkspaceImage()

Deletes the specified image from your account. To delete an image, you must first delete any bundles that are associated with the image and unshare the image if it is shared with other accounts.

Required Parameters

ImageId: The identifier of the image.

Main.Workspaces.deregister_workspace_directory — Method

DeregisterWorkspaceDirectory()

Deregisters the specified directory. This operation is asynchronous and returns before the WorkSpace directory is deregistered. If any WorkSpaces are registered to this directory, you must remove them before you can deregister the directory. Simple AD and AD Connector are made available to you free of charge to use with WorkSpaces. If there are no WorkSpaces being used with your Simple AD or AD Connector directory for 30 consecutive days, this directory will be automatically deregistered for use with Amazon WorkSpaces, and you will be charged for this directory as per the AWS Directory Services pricing terms. To delete empty directories, see Delete the Directory for Your WorkSpaces. If you delete your Simple AD or AD Connector directory, you can always create a new one when you want to start using WorkSpaces again.

Required Parameters

DirectoryId: The identifier of the directory. If any WorkSpaces are registered to this directory, you must remove them before you deregister the directory, or you will receive an OperationNotSupportedException error.

Main.Workspaces.describe_account — Method

DescribeAccount()

Retrieves a list that describes the configuration of Bring Your Own License (BYOL) for the specified account.

Main.Workspaces.describe_account_modifications — Method

DescribeAccountModifications()

Retrieves a list that describes modifications to the configuration of Bring Your Own License (BYOL) for the specified account.

Optional Parameters

NextToken: If you received a NextToken from a previous call that was paginated, provide this token to receive the next set of results.

Main.Workspaces.describe_client_properties — Method

DescribeClientProperties()

Retrieves a list that describes one or more specified Amazon WorkSpaces clients.

Required Parameters

ResourceIds: The resource identifier, in the form of directory IDs.

Main.Workspaces.describe_connection_alias_permissions — Method

DescribeConnectionAliasPermissions()

Describes the permissions that the owner of a connection alias has granted to another AWS account for the specified connection alias. For more information, see Cross-Region Redirection for Amazon WorkSpaces.

Required Parameters

AliasId: The identifier of the connection alias.

Optional Parameters

MaxResults: The maximum number of results to return.

NextToken: If you received a NextToken from a previous call that was paginated, provide this token to receive the next set of results.

Main.Workspaces.describe_connection_aliases — Method

DescribeConnectionAliases()

Retrieves a list that describes the connection aliases used for cross-Region redirection. For more information, see Cross-Region Redirection for Amazon WorkSpaces.

Optional Parameters

AliasIds: The identifiers of the connection aliases to describe.

Limit: The maximum number of connection aliases to return.

NextToken: If you received a NextToken from a previous call that was paginated, provide this token to receive the next set of results.

ResourceId: The identifier of the directory associated with the connection alias.

Main.Workspaces.describe_ip_groups — Method

DescribeIpGroups()

Describes one or more of your IP access control groups.

Optional Parameters

GroupIds: The identifiers of one or more IP access control groups.

MaxResults: The maximum number of items to return.

NextToken: If you received a NextToken from a previous call that was paginated, provide this token to receive the next set of results.

Main.Workspaces.describe_tags — Method

DescribeTags()

Describes the specified tags for the specified WorkSpaces resource.

Required Parameters

ResourceId: The identifier of the WorkSpaces resource. The supported resource types are WorkSpaces, registered directories, images, custom bundles, IP access control groups, and connection aliases.

Main.Workspaces.describe_workspace_bundles — Method

DescribeWorkspaceBundles()

Retrieves a list that describes the available WorkSpace bundles. You can filter the results using either bundle ID or owner, but not both.

Optional Parameters

BundleIds: The identifiers of the bundles. You cannot combine this parameter with any other filter.

NextToken: The token for the next set of results. (You received this token from a previous call.)

Owner: The owner of the bundles. You cannot combine this parameter with any other filter. Specify AMAZON to describe the bundles provided by AWS or null to describe the bundles that belong to your account.

Main.Workspaces.describe_workspace_directories — Method

DescribeWorkspaceDirectories()

Describes the available directories that are registered with Amazon WorkSpaces.

Optional Parameters

DirectoryIds: The identifiers of the directories. If the value is null, all directories are retrieved.

Limit: The maximum number of directories to return.

NextToken: If you received a NextToken from a previous call that was paginated, provide this token to receive the next set of results.

Main.Workspaces.describe_workspace_image_permissions — Method

DescribeWorkspaceImagePermissions()

Describes the permissions that the owner of an image has granted to other AWS accounts for an image.

Required Parameters

ImageId: The identifier of the image.

Optional Parameters

MaxResults: The maximum number of items to return.

NextToken: If you received a NextToken from a previous call that was paginated, provide this token to receive the next set of results.

Main.Workspaces.describe_workspace_images — Method

DescribeWorkspaceImages()

Retrieves a list that describes one or more specified images, if the image identifiers are provided. Otherwise, all images in the account are described.

Optional Parameters

ImageIds: The identifier of the image.

ImageType: The type (owned or shared) of the image.

MaxResults: The maximum number of items to return.

NextToken: If you received a NextToken from a previous call that was paginated, provide this token to receive the next set of results.

Main.Workspaces.describe_workspace_snapshots — Method

DescribeWorkspaceSnapshots()

Describes the snapshots for the specified WorkSpace.

Required Parameters

WorkspaceId: The identifier of the WorkSpace.

Main.Workspaces.describe_workspaces — Method

DescribeWorkspaces()

Describes the specified WorkSpaces. You can filter the results by using the bundle identifier, directory identifier, or owner, but you can specify only one filter at a time.

Optional Parameters

BundleId: The identifier of the bundle. All WorkSpaces that are created from this bundle are retrieved. You cannot combine this parameter with any other filter.

DirectoryId: The identifier of the directory. In addition, you can optionally specify a specific directory user (see UserName). You cannot combine this parameter with any other filter.

Limit: The maximum number of items to return.

NextToken: If you received a NextToken from a previous call that was paginated, provide this token to receive the next set of results.

UserName: The name of the directory user. You must specify this parameter with DirectoryId.

WorkspaceIds: The identifiers of the WorkSpaces. You cannot combine this parameter with any other filter. Because the CreateWorkspaces operation is asynchronous, the identifier it returns is not immediately available. If you immediately call DescribeWorkspaces with this identifier, no information is returned.

Main.Workspaces.describe_workspaces_connection_status — Method

DescribeWorkspacesConnectionStatus()

Describes the connection status of the specified WorkSpaces.

Optional Parameters

NextToken: If you received a NextToken from a previous call that was paginated, provide this token to receive the next set of results.

WorkspaceIds: The identifiers of the WorkSpaces. You can specify up to 25 WorkSpaces.

Main.Workspaces.disassociate_connection_alias — Method

DisassociateConnectionAlias()

Disassociates a connection alias from a directory. Disassociating a connection alias disables cross-Region redirection between two directories in different AWS Regions. For more information, see Cross-Region Redirection for Amazon WorkSpaces. Before performing this operation, call DescribeConnectionAliases to make sure that the current state of the connection alias is CREATED.

Required Parameters

AliasId: The identifier of the connection alias to disassociate.

Main.Workspaces.disassociate_ip_groups — Method

DisassociateIpGroups()

Disassociates the specified IP access control group from the specified directory.

Required Parameters

DirectoryId: The identifier of the directory.

GroupIds: The identifiers of one or more IP access control groups.

Main.Workspaces.import_workspace_image — Method

ImportWorkspaceImage()

Imports the specified Windows 10 Bring Your Own License (BYOL) image into Amazon WorkSpaces. The image must be an already licensed Amazon EC2 image that is in your AWS account, and you must own the image. For more information about creating BYOL images, see Bring Your Own Windows Desktop Licenses.

Required Parameters

Ec2ImageId: The identifier of the EC2 image.

ImageDescription: The description of the WorkSpace image.

ImageName: The name of the WorkSpace image.

IngestionProcess: The ingestion process to be used when importing the image, depending on which protocol you want to use for your BYOL Workspace image, either PCoIP or WorkSpaces Streaming Protocol (WSP). To use WSP, specify a value that ends in WSP. To use PCoIP, specify a value that does not end in _WSP. For non-GPU-enabled bundles (bundles other than Graphics or GraphicsPro), specify BYOLREGULAR or BYOLREGULARWSP, depending on the protocol.

Optional Parameters

Applications: If specified, the version of Microsoft Office to subscribe to. Valid only for Windows 10 BYOL images. For more information about subscribing to Office for BYOL images, see Bring Your Own Windows Desktop Licenses. Although this parameter is an array, only one item is allowed at this time.

Tags: The tags. Each WorkSpaces resource can have a maximum of 50 tags.

Main.Workspaces.list_available_management_cidr_ranges — Method

ListAvailableManagementCidrRanges()

Retrieves a list of IP address ranges, specified as IPv4 CIDR blocks, that you can use for the network management interface when you enable Bring Your Own License (BYOL). This operation can be run only by AWS accounts that are enabled for BYOL. If your account isn't enabled for BYOL, you'll receive an AccessDeniedException error. The management network interface is connected to a secure Amazon WorkSpaces management network. It is used for interactive streaming of the WorkSpace desktop to Amazon WorkSpaces clients, and to allow Amazon WorkSpaces to manage the WorkSpace.

Required Parameters

ManagementCidrRangeConstraint: The IP address range to search. Specify an IP address range that is compatible with your network and in CIDR notation (that is, specify the range as an IPv4 CIDR block).

Optional Parameters

MaxResults: The maximum number of items to return.

NextToken: If you received a NextToken from a previous call that was paginated, provide this token to receive the next set of results.

Main.Workspaces.migrate_workspace — Method

MigrateWorkspace()

Migrates a WorkSpace from one operating system or bundle type to another, while retaining the data on the user volume. The migration process recreates the WorkSpace by using a new root volume from the target bundle image and the user volume from the last available snapshot of the original WorkSpace. During migration, the original D:Users%USERNAME% user profile folder is renamed to D:Users%USERNAME%MMddyyTHHmmss%.NotMigrated. A new D:Users%USERNAME% folder is generated by the new OS. Certain files in the old user profile are moved to the new user profile. For available migration scenarios, details about what happens during migration, and best practices, see Migrate a WorkSpace.

Required Parameters

BundleId: The identifier of the target bundle type to migrate the WorkSpace to.

SourceWorkspaceId: The identifier of the WorkSpace to migrate from.

Main.Workspaces.modify_account — Method

ModifyAccount()

Modifies the configuration of Bring Your Own License (BYOL) for the specified account.

Optional Parameters

DedicatedTenancyManagementCidrRange: The IP address range, specified as an IPv4 CIDR block, for the management network interface. Specify an IP address range that is compatible with your network and in CIDR notation (that is, specify the range as an IPv4 CIDR block). The CIDR block size must be /16 (for example, 203.0.113.25/16). It must also be specified as available by the ListAvailableManagementCidrRanges operation.

DedicatedTenancySupport: The status of BYOL.

Main.Workspaces.modify_client_properties — Method

ModifyClientProperties()

Modifies the properties of the specified Amazon WorkSpaces clients.

Required Parameters

ClientProperties: Information about the Amazon WorkSpaces client.

ResourceId: The resource identifiers, in the form of directory IDs.

Main.Workspaces.modify_selfservice_permissions — Method

ModifySelfservicePermissions()

Modifies the self-service WorkSpace management capabilities for your users. For more information, see Enable Self-Service WorkSpace Management Capabilities for Your Users.

Required Parameters

ResourceId: The identifier of the directory.

SelfservicePermissions: The permissions to enable or disable self-service capabilities.

Main.Workspaces.modify_workspace_access_properties — Method

ModifyWorkspaceAccessProperties()

Specifies which devices and operating systems users can use to access their WorkSpaces. For more information, see Control Device Access.

Required Parameters

ResourceId: The identifier of the directory.

WorkspaceAccessProperties: The device types and operating systems to enable or disable for access.

Main.Workspaces.modify_workspace_creation_properties — Method

ModifyWorkspaceCreationProperties()

Modify the default properties used to create WorkSpaces.

Required Parameters

ResourceId: The identifier of the directory.

WorkspaceCreationProperties: The default properties for creating WorkSpaces.

Main.Workspaces.modify_workspace_properties — Method

ModifyWorkspaceProperties()

Modifies the specified WorkSpace properties. For important information about how to modify the size of the root and user volumes, see Modify a WorkSpace.

Required Parameters

WorkspaceId: The identifier of the WorkSpace.

WorkspaceProperties: The properties of the WorkSpace.

Main.Workspaces.modify_workspace_state — Method

ModifyWorkspaceState()

Sets the state of the specified WorkSpace. To maintain a WorkSpace without being interrupted, set the WorkSpace state to ADMINMAINTENANCE. WorkSpaces in this state do not respond to requests to reboot, stop, start, rebuild, or restore. An AutoStop WorkSpace in this state is not stopped. Users cannot log into a WorkSpace in the ADMINMAINTENANCE state.

Required Parameters

WorkspaceId: The identifier of the WorkSpace.

WorkspaceState: The WorkSpace state.

Main.Workspaces.reboot_workspaces — Method

RebootWorkspaces()

Reboots the specified WorkSpaces. You cannot reboot a WorkSpace unless its state is AVAILABLE or UNHEALTHY. This operation is asynchronous and returns before the WorkSpaces have rebooted.

Required Parameters

RebootWorkspaceRequests: The WorkSpaces to reboot. You can specify up to 25 WorkSpaces.

Main.Workspaces.rebuild_workspaces — Method

RebuildWorkspaces()

Rebuilds the specified WorkSpace. You cannot rebuild a WorkSpace unless its state is AVAILABLE, ERROR, UNHEALTHY, STOPPED, or REBOOTING. Rebuilding a WorkSpace is a potentially destructive action that can result in the loss of data. For more information, see Rebuild a WorkSpace. This operation is asynchronous and returns before the WorkSpaces have been completely rebuilt.

Required Parameters

RebuildWorkspaceRequests: The WorkSpace to rebuild. You can specify a single WorkSpace.

Main.Workspaces.register_workspace_directory — Method

RegisterWorkspaceDirectory()

Registers the specified directory. This operation is asynchronous and returns before the WorkSpace directory is registered. If this is the first time you are registering a directory, you will need to create the workspacesDefaultRole role before you can register a directory. For more information, see Creating the workspacesDefaultRole Role.

Required Parameters

DirectoryId: The identifier of the directory. You cannot register a directory if it does not have a status of Active. If the directory does not have a status of Active, you will receive an InvalidResourceStateException error. If you have already registered the maximum number of directories that you can register with Amazon WorkSpaces, you will receive a ResourceLimitExceededException error. Deregister directories that you are not using for WorkSpaces, and try again.

EnableWorkDocs: Indicates whether Amazon WorkDocs is enabled or disabled. If you have enabled this parameter and WorkDocs is not available in the Region, you will receive an OperationNotSupportedException error. Set EnableWorkDocs to disabled, and try again.

Optional Parameters

EnableSelfService: Indicates whether self-service capabilities are enabled or disabled.

SubnetIds: The identifiers of the subnets for your virtual private cloud (VPC). Make sure that the subnets are in supported Availability Zones. The subnets must also be in separate Availability Zones. If these conditions are not met, you will receive an OperationNotSupportedException error.

Tags: The tags associated with the directory.

Tenancy: Indicates whether your WorkSpace directory is dedicated or shared. To use Bring Your Own License (BYOL) images, this value must be set to DEDICATED and your AWS account must be enabled for BYOL. If your account has not been enabled for BYOL, you will receive an InvalidParameterValuesException error. For more information about BYOL images, see Bring Your Own Windows Desktop Images.

Main.Workspaces.restore_workspace — Method

RestoreWorkspace()

Restores the specified WorkSpace to its last known healthy state. You cannot restore a WorkSpace unless its state is AVAILABLE, ERROR, UNHEALTHY, or STOPPED. Restoring a WorkSpace is a potentially destructive action that can result in the loss of data. For more information, see Restore a WorkSpace. This operation is asynchronous and returns before the WorkSpace is completely restored.

Required Parameters

WorkspaceId: The identifier of the WorkSpace.

Main.Workspaces.revoke_ip_rules — Method

RevokeIpRules()

Removes one or more rules from the specified IP access control group.

Required Parameters

GroupId: The identifier of the group.

UserRules: The rules to remove from the group.

Main.Workspaces.start_workspaces — Method

StartWorkspaces()

Starts the specified WorkSpaces. You cannot start a WorkSpace unless it has a running mode of AutoStop and a state of STOPPED.

Required Parameters

StartWorkspaceRequests: The WorkSpaces to start. You can specify up to 25 WorkSpaces.

Main.Workspaces.stop_workspaces — Method

StopWorkspaces()

Stops the specified WorkSpaces. You cannot stop a WorkSpace unless it has a running mode of AutoStop and a state of AVAILABLE, IMPAIRED, UNHEALTHY, or ERROR.

Required Parameters

StopWorkspaceRequests: The WorkSpaces to stop. You can specify up to 25 WorkSpaces.

Main.Workspaces.terminate_workspaces — Method

TerminateWorkspaces()

Terminates the specified WorkSpaces. Terminating a WorkSpace is a permanent action and cannot be undone. The user's data is destroyed. If you need to archive any user data, contact AWS Support before terminating the WorkSpace. You can terminate a WorkSpace that is in any state except SUSPENDED. This operation is asynchronous and returns before the WorkSpaces have been completely terminated. After a WorkSpace is terminated, the TERMINATED state is returned only briefly before the WorkSpace directory metadata is cleaned up, so this state is rarely returned. To confirm that a WorkSpace is terminated, check for the WorkSpace ID by using DescribeWorkSpaces. If the WorkSpace ID isn't returned, then the WorkSpace has been successfully terminated. Simple AD and AD Connector are made available to you free of charge to use with WorkSpaces. If there are no WorkSpaces being used with your Simple AD or AD Connector directory for 30 consecutive days, this directory will be automatically deregistered for use with Amazon WorkSpaces, and you will be charged for this directory as per the AWS Directory Services pricing terms. To delete empty directories, see Delete the Directory for Your WorkSpaces. If you delete your Simple AD or AD Connector directory, you can always create a new one when you want to start using WorkSpaces again.

Required Parameters

TerminateWorkspaceRequests: The WorkSpaces to terminate. You can specify up to 25 WorkSpaces.

Main.Workspaces.update_connection_alias_permission — Method

UpdateConnectionAliasPermission()

Shares or unshares a connection alias with one account by specifying whether that account has permission to associate the connection alias with a directory. If the association permission is granted, the connection alias is shared with that account. If the association permission is revoked, the connection alias is unshared with the account. For more information, see Cross-Region Redirection for Amazon WorkSpaces. Before performing this operation, call DescribeConnectionAliases to make sure that the current state of the connection alias is CREATED. To delete a connection alias that has been shared, the shared account must first disassociate the connection alias from any directories it has been associated with. Then you must unshare the connection alias from the account it has been shared with. You can delete a connection alias only after it is no longer shared with any accounts or associated with any directories.

Required Parameters

AliasId: The identifier of the connection alias that you want to update permissions for.

ConnectionAliasPermission: Indicates whether to share or unshare the connection alias with the specified AWS account.

Main.Workspaces.update_rules_of_ip_group — Method

UpdateRulesOfIpGroup()

Replaces the current rules of the specified IP access control group with the specified rules.

Required Parameters

GroupId: The identifier of the group.

UserRules: One or more rules.

Main.Workspaces.update_workspace_image_permission — Method

UpdateWorkspaceImagePermission()

Shares or unshares an image with one account in the same AWS Region by specifying whether that account has permission to copy the image. If the copy image permission is granted, the image is shared with that account. If the copy image permission is revoked, the image is unshared with the account. After an image has been shared, the recipient account can copy the image to other AWS Regions as needed. In the China (Ningxia) Region, you can copy images only within the same Region. In the AWS GovCloud (US-West) Region, to copy images to and from other AWS Regions, contact AWS Support. For more information about sharing images, see Share or Unshare a Custom WorkSpaces Image. To delete an image that has been shared, you must unshare the image before you delete it. Sharing Bring Your Own License (BYOL) images across AWS accounts isn't supported at this time in the AWS GovCloud (US-West) Region. To share BYOL images across accounts in the AWS GovCloud (US-West) Region, contact AWS Support.

Required Parameters

AllowCopyImage: The permission to copy the image. This permission can be revoked only after an image has been shared.

ImageId: The identifier of the image.

SharedAccountId: The identifier of the AWS account to share or unshare the image with. Before sharing the image, confirm that you are sharing to the correct AWS account ID.